% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_106.R
\name{memlit}
\alias{memlit}
\title{Membership in literary or art group}
\source{
General Social Survey https://gss.norc.org
}
\description{
memlit
}
\details{
Question Now we would like to know something about the groups or organizations to which individuals belong. Here is a list of various organizations. Could you tell me whether or not you are a member of each type?  READ EACH ITEM. CODE OnE FOR EACH.
M. Literary, art, discussion or study groups
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1974 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab B/C/- \tab partial \cr
   1989 \tab B/C/- \tab partial \cr
   1990 \tab B/C/- \tab partial \cr
   1991 \tab B/C/- \tab partial \cr
   1993 \tab B/C/- \tab partial \cr
   1994 \tab B/C/- \tab partial \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/975/vshow}{GSS Data Explorer page for \code{memlit}}.
}

\seealso{
Other Socio-political participation: 
\code{\link{memchurh}},
\code{\link{memfarm}},
\code{\link{memgreek}},
\code{\link{memhobby}},
\code{\link{memnat}},
\code{\link{memnum}},
\code{\link{memother}},
\code{\link{memprof}},
\code{\link{memschl}}
}
\concept{Socio-political participation}
\keyword{variable}
