% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_105.R
\name{mempolit}
\alias{mempolit}
\title{Membership in political club}
\source{
General Social Survey https://gss.norc.org
}
\description{
mempolit
}
\details{
Question Now we would like to know something about the groups or organizations to which individuals belong. Here is a list of various organizations. Could you tell me whether or not you are a member of each type?  READ EACH ITEM. CODE OnE FOR EACH.
D. Political clubs
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1974 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab B/C/- \tab partial \cr
   1989 \tab B/C/- \tab partial \cr
   1990 \tab B/C/- \tab partial \cr
   1991 \tab B/C/- \tab partial \cr
   1993 \tab B/C/- \tab partial \cr
   1994 \tab B/C/- \tab partial \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/966/vshow}{GSS Data Explorer page for \code{mempolit}}.
}

\seealso{
Other Socio-political participation: 
\code{\link{actchurh}},
\code{\link{actfarm}},
\code{\link{actfrat}},
\code{\link{actgreek}},
\code{\link{acthobby}},
\code{\link{actlit}},
\code{\link{actnat}},
\code{\link{actother}},
\code{\link{actpolit}},
\code{\link{actprof}},
\code{\link{actschl}},
\code{\link{actserv}},
\code{\link{actsport}},
\code{\link{actunion}},
\code{\link{actvet}},
\code{\link{actyouth}},
\code{\link{attndgrp}},
\code{\link{blkgains}},
\code{\link{blkinflu}},
\code{\link{caregrp}},
\code{\link{churhgrp}},
\code{\link{fedtrust}},
\code{\link{fegains}},
\code{\link{feinflu}},
\code{\link{gavepol}},
\code{\link{givegrp}},
\code{\link{governor}},
\code{\link{intpol}},
\code{\link{leadgrp}},
\code{\link{lobbygrp}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{locinflu}},
\code{\link{loclived}},
\code{\link{loclobby}},
\code{\link{locprob}},
\code{\link{locself}},
\code{\link{loctrust}},
\code{\link{locvote}},
\code{\link{memchurh}},
\code{\link{memfarm}},
\code{\link{memfrat}},
\code{\link{memgreek}},
\code{\link{memhobby}},
\code{\link{memlit}},
\code{\link{memnat}},
\code{\link{memnum}},
\code{\link{memother}},
\code{\link{memprof}},
\code{\link{memschl}},
\code{\link{memserv}},
\code{\link{memsport}},
\code{\link{memunion}},
\code{\link{memvet}},
\code{\link{memyouth}},
\code{\link{mostact}},
\code{\link{othlobby}},
\code{\link{othself}},
\code{\link{polrally}},
\code{\link{schlhead}},
\code{\link{servegrp}},
\code{\link{solchurh}},
\code{\link{solfarm}},
\code{\link{solfrat}},
\code{\link{solgreek}},
\code{\link{solhobby}},
\code{\link{sollit}},
\code{\link{solnat}},
\code{\link{solother}},
\code{\link{solpolit}},
\code{\link{solprof}},
\code{\link{solschl}},
\code{\link{solserv}},
\code{\link{solsport}},
\code{\link{solunion}},
\code{\link{solvet}},
\code{\link{solyouth}},
\code{\link{swayvote}},
\code{\link{usrep}},
\code{\link{workpol}},
\code{\link{writegrp}}
}
\concept{Socio-political participation}
\keyword{variable}
