% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_182.R
\name{men}
\alias{men}
\title{Men's lives improved}
\source{
General Social Survey https://gss.norc.org
}
\description{
men
}
\details{
Question We'd like your views on how the women's movement has affected certain groups. For each group I name, please tell me whether you think the women's movement has improved their lives, made their lives worse, or had no effect On their lives.
D. Men
}
\section{Values}{

\itemize{
\item \code{1} improved their life
\item \code{2} made their life worse
\item \code{3} had no effect on their life
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1703/vshow}{GSS Data Explorer page for \code{men}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab improved their life \tab made their life worse \tab had no effect on their life \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 427 \tab 431 \tab 485 \tab 113 \tab 4 \tab 2904 \cr
   Total \tab 1444 \tab 427 \tab 431 \tab 485 \tab 113 \tab 4 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{children}},
\code{\link{manprof}},
\code{\link{yourself}}
}
\concept{Gender}
\keyword{variable}
