% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_483.R
\name{meovrwrk}
\alias{meovrwrk}
\title{Men hurt family when focus on work too much}
\source{
General Social Survey https://gss.norc.org
}
\description{
meovrwrk
}
\details{
Question And, do you agree or disagree . . .
C. Family life often suffers because men concentrate too much on their work.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   2000 \tab A/B/- \tab partial \cr
   2002 \tab A/B/- \tab partial \cr
   2004 \tab A/B/- \tab partial \cr
   2006 \tab A/B/- \tab partial \cr
   2008 \tab A/B/- \tab partial \cr
   2010 \tab A/B/- \tab partial \cr
   2012 \tab A/B/- \tab partial \cr
   2014 \tab A/B/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
   2021 \tab A/B/- \tab partial \cr
   2022 \tab A/B/- \tab partial \cr
   2024 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4404/vshow}{GSS Data Explorer page for \code{meovrwrk}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{cohabfst}},
\code{\link{cohabok}},
\code{\link{divbest}},
\code{\link{divifkd1}},
\code{\link{divnokd1}},
\code{\link{fewrksup}},
\code{\link{hubbywk1}},
\code{\link{mrmom}},
\code{\link{singlpar}}

Other Men: 
\code{\link{hubbywk1}},
\code{\link{mrmom}}

Other Women: 
\code{\link{fewrksup}},
\code{\link{hubbywk1}},
\code{\link{mrmom}}
}
\concept{Family}
\concept{Gender}
\concept{ISSP Family & Gender Roles}
\concept{Men}
\concept{Women}
\keyword{variable}
