% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_135.R
\name{merged}
\alias{merged}
\title{The org. was merged, taken over}
\source{
General Social Survey https://gss.norc.org
}
\description{
merged
}
\details{
Question In the last five years, did any of the following things happen to an organization that you worked for?
B. The organization was merged, taken over, or changed ownership?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1268/vshow}{GSS Data Explorer page for \code{merged}}.
}

\seealso{
Other Work Organizations: 
\code{\link{genecanx}},
\code{\link{genecany}},
\code{\link{genegets}},
\code{\link{outofbiz}},
\code{\link{profits1}},
\code{\link{profits2}},
\code{\link{reorg}},
\code{\link{unpower}},
\code{\link{unprog}}

Other Corporations: 
\code{\link{outofbiz}},
\code{\link{profits1}},
\code{\link{profits2}},
\code{\link{reorg}}
}
\concept{Corporations}
\concept{Work Organizations}
\keyword{variable}
