% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_416.R
\name{mhcause}
\alias{mhcause}
\title{What caused x s situation}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhcause
}
\details{
Question Would you say that (NAME)'s situation is caused by depression, asthma, schizophrenia, stress, or something else?
}
\section{Values}{

\itemize{
\item \code{1} depression
\item \code{2} asthma
\item \code{3} schizophrenia
\item \code{4} stress
\item \code{5} something else
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3801/vshow}{GSS Data Explorer page for \code{mhcause}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab depression \tab asthma \tab schizophrenia \tab stress \tab something else \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 551 \tab 304 \tab 233 \tab 132 \tab 124 \tab 71 \tab 10 \tab 4510 \cr
   Total \tab 3085 \tab 551 \tab 304 \tab 233 \tab 132 \tab 124 \tab 71 \tab 10 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhclsoth}},
\code{\link{mhdisoth}},
\code{\link{mhhlpoth}},
\code{\link{mhresoth}},
\code{\link{mhseroth}},
\code{\link{mhtrtot2}},
\code{\link{mhviooth}},
\code{\link{mhvioslf}}
}
\concept{Mental Health IV}
\keyword{variable}
