% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_412.R
\name{mhdecslf}
\alias{mhdecslf}
\title{How able is x to decide whether to receive treatment}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhdecslf
}
\details{
Question How able is (NAME) to decide whether or not (he/she) should receive treatment. Would you say…
}
\section{Values}{

\itemize{
\item \code{1} very able
\item \code{2} moderately able
\item \code{3} not very able
\item \code{4} not at all able
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3760/vshow}{GSS Data Explorer page for \code{mhdecslf}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very able \tab moderately able \tab not very able \tab not at all able \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 389 \tab 488 \tab 407 \tab 103 \tab 30 \tab 8 \tab 4510 \cr
   Total \tab 3085 \tab 389 \tab 488 \tab 407 \tab 103 \tab 30 \tab 8 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhimptrt}},
\code{\link{mhintl}},
\code{\link{mhlessop}},
\code{\link{mhlosefr}},
\code{\link{mhmoney}},
\code{\link{mhnotpol}},
\code{\link{mhoutsdr}},
\code{\link{mhuneasy}},
\code{\link{mhunsure}}
}
\concept{Mental Health IV}
\keyword{variable}
