% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_416.R
\name{mhdiagno}
\alias{mhdiagno}
\title{R knows someone who has been diagnosed with a mental health problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhdiagno
}
\details{
Question Leaving yourself aside, have you ever known someone personally who has been diagnosed with a mental health problem?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
   D/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7228/vshow}{GSS Data Explorer page for \code{mhdiagno}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2018 \tab 1175 \tab 597 \tab 533 \tab 6 \tab 37 \tab 2348 \cr
   Total \tab 5685 \tab 597 \tab 533 \tab 6 \tab 37 \tab 6858 \cr
}
}

\keyword{variable}
