% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_416.R
\name{mhdisoth}
\alias{mhdisoth}
\title{How much distress this person caused r}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhdisoth
}
\details{
Question Let's turn away from (NAME) now. I have a few questions about your general experiences that I would like to ask you. Leaving yourself aside, have you personally ever known someone who has received treatment for a mental health situation?
E. How much distress did this person's mental health problem cause you?
}
\section{Values}{

\itemize{
\item \verb{[1]} a great deal
\item \verb{[2]} quite a bit
\item \verb{[3]} a little
\item \verb{[4]} not at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3807/vshow}{GSS Data Explorer page for \code{mhdisoth}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab quite a bit \tab a little \tab not at all \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3535 \tab 148 \tab 204 \tab 340 \tab 275 \tab 1 \tab 7 \tab 4510 \cr
   Total \tab 3535 \tab 148 \tab 204 \tab 340 \tab 275 \tab 1 \tab 7 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhcause}},
\code{\link{mhclsoth}},
\code{\link{mhhlpoth}},
\code{\link{mhresoth}},
\code{\link{mhseroth}},
\code{\link{mhtrtot2}},
\code{\link{mhviooth}},
\code{\link{mhvioslf}}
}
\concept{Mental Health IV}
\keyword{variable}
