% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_411.R
\name{mhfriend}
\alias{mhfriend}
\title{R would make friends with x}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhfriend
}
\details{
Question Would you be definitely willing, probably willing, probably unwilling, or definitely unwilling…
D. To make friends with (NAME)?
}
\section{Values}{

\itemize{
\item \code{1} definitely willing
\item \code{2} probably willing
\item \code{3} probably unwilling
\item \code{4} definitely unwilling
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3755/vshow}{GSS Data Explorer page for \code{mhfriend}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely willing \tab probably willing \tab probably unwilling \tab definitely unwilling \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 410 \tab 776 \tab 160 \tab 55 \tab 18 \tab 6 \tab 4510 \cr
   Total \tab 3085 \tab 410 \tab 776 \tab 160 \tab 55 \tab 18 \tab 6 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhillnss}},
\code{\link{mhimpown}},
\code{\link{mhkdcare}},
\code{\link{mhneihbr}},
\code{\link{mhphyscl}},
\code{\link{mhsocial}},
\code{\link{mhupdown}},
\code{\link{mhwedrel}},
\code{\link{mhwkwith}}
}
\concept{Mental Health IV}
\keyword{variable}
