% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_417.R
\name{mhfright}
\alias{mhfright}
\title{How frightening most people are in public places}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhfright
}
\details{
Question Do you frequently, occasionally, rarely, or never see someone who seems to have a serious mental health problem in a public place?
A. How frightening are most people that you see in public places who appear to have a serious mental health problem?  Are they…
}
\section{Values}{

\itemize{
\item \verb{[1]} very frightening
\item \verb{[2]} somewhat frightening
\item \verb{[3]} not very frightening
\item \verb{[4]} not at all frightening
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3812/vshow}{GSS Data Explorer page for \code{mhfright}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very frightening \tab somewhat frightening \tab not very frightening \tab not at all frightening \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3343 \tab 29 \tab 324 \tab 490 \tab 305 \tab 13 \tab 6 \tab 4510 \cr
   Total \tab 3343 \tab 29 \tab 324 \tab 490 \tab 305 \tab 13 \tab 6 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhexpoth}},
\code{\link{mhreloth}},
\code{\link{mhseepub}},
\code{\link{mhsymp}},
\code{\link{mhtrtslf}}
}
\concept{Mental Health IV}
\keyword{variable}
