% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_410.R
\name{mhgod}
\alias{mhgod}
\title{X s situation caused by: god s will}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhgod
}
\details{
Question Is it very likely, somewhat likely, not very likely, or not at all likely that (NAME)''s situation is caused by…
F. God''s will
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3747/vshow}{GSS Data Explorer page for \code{mhgod}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 96 \tab 190 \tab 305 \tab 796 \tab 29 \tab 9 \tab 4510 \cr
   Total \tab 3085 \tab 96 \tab 190 \tab 305 \tab 796 \tab 29 \tab 9 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhbrain}},
\code{\link{mhchrctr}},
\code{\link{mhdomove}},
\code{\link{mhdonone}},
\code{\link{mhdooth}},
\code{\link{mhgenes}},
\code{\link{mhluck}},
\code{\link{mhraised}},
\code{\link{mhstress}}
}
\concept{Mental Health IV}
\keyword{variable}
