% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_415.R
\name{mhgvthlt}
\alias{mhgvthlt}
\title{Govt responsibility: provide health care for x}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhgvthlt
}
\details{
Question The next few questions deal with the government's responsibility to help people like (NAME). For each statement please tell me if you think the government definitely should, probably should, probably should not, or definitely should not be responsible for helping people with situations like (NAME).
B. Provide health care for (NAME).
}
\section{Values}{

\itemize{
\item \code{1} definitely should be
\item \code{2} probably should be
\item \code{3} probably should not be
\item \code{4} definitely should not be
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3789/vshow}{GSS Data Explorer page for \code{mhgvthlt}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should be \tab probably should be \tab probably should not be \tab definitely should not be \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 488 \tab 617 \tab 200 \tab 77 \tab 33 \tab 10 \tab 4510 \cr
   Total \tab 3085 \tab 488 \tab 617 \tab 200 \tab 77 \tab 33 \tab 10 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhfamshm}},
\code{\link{mhgvtdis}},
\code{\link{mhgvtedc}},
\code{\link{mhgvthme}},
\code{\link{mhhsptrt}},
\code{\link{mhmeds}},
\code{\link{mhmnthsp}},
\code{\link{mhseedoc}},
\code{\link{mhslfshm}}
}
\concept{Mental Health IV}
\keyword{variable}
