% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_412.R
\name{mhimptrt}
\alias{mhimptrt}
\title{How likely x s situation will improve with treatment}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhimptrt
}
\details{
Question How likely is it that (NAME)'s situation will improve with treatment? Is it very likely, somewhat likely, not very likely, or not at all likely?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3759/vshow}{GSS Data Explorer page for \code{mhimptrt}}.
}

\seealso{
Other Mental Health IV: 
\code{\link{mhdecslf}},
\code{\link{mhintl}},
\code{\link{mhlessop}},
\code{\link{mhlosefr}},
\code{\link{mhmoney}},
\code{\link{mhnotpol}},
\code{\link{mhoutsdr}},
\code{\link{mhuneasy}},
\code{\link{mhunsure}}
}
\concept{Mental Health IV}
\keyword{variable}
