% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_413.R
\name{mhnervs}
\alias{mhnervs}
\title{Being around x would make me feel nervous}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhnervs
}
\details{
Question For the next several questions, please tell me whether you strongly agree, agree, disagree, or strongly disagree with the statement.
K. Being around (NAME) would make me feel nervous.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3772/vshow}{GSS Data Explorer page for \code{mhnervs}}.
}

\seealso{
Other Mental Health IV: 
\code{\link{mhaccptd}},
\code{\link{mhcreatv}},
\code{\link{mhfeelem}},
\code{\link{mhhired}},
\code{\link{mhhrdtlk}},
\code{\link{mhnokids}},
\code{\link{mhnotell}},
\code{\link{mhprdctv}},
\code{\link{mhtrusty}}

Other Mental Health: 
\code{\link{mhaccptd}},
\code{\link{mhcreatv}},
\code{\link{mhfeelem}},
\code{\link{mhhired}},
\code{\link{mhhrdtlk}},
\code{\link{mhnokids}},
\code{\link{mhnotell}},
\code{\link{mhprdctv}},
\code{\link{mhtrusty}}
}
\concept{Mental Health}
\concept{Mental Health IV}
\keyword{variable}
