% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_392.R
\name{mhothrel}
\alias{mhothrel}
\title{Result of mental health problem to relationship with him}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhothrel
}
\details{
Question As a result of the mental health problem, would you say that your relationship with this person…
}
\section{Values}{

\itemize{
\item \code{1} became stronger
\item \code{2} became worse
\item \code{3} was unchanged
\item \code{4} ended as a result of the problem
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3567/vshow}{GSS Data Explorer page for \code{mhothrel}}.
}

\seealso{
Other Mental Health III: 
\code{\link{mhothyou}},
\code{\link{mhtrtoth}},
\code{\link{seemhpub}},
\code{\link{viglab06}}
}
\concept{Mental Health III}
\keyword{variable}
