% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_392.R
\name{mhothyou}
\alias{mhothyou}
\title{How much distress his mental health problem caused r}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhothyou
}
\details{
Question How much distress did this person’s mental health problem cause you…
}
\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} quite a bit
\item \code{3} a little
\item \code{4} not at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3566/vshow}{GSS Data Explorer page for \code{mhothyou}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab quite a bit \tab a little \tab not at all \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3595 \tab 240 \tab 179 \tab 281 \tab 200 \tab 1 \tab 14 \tab 4510 \cr
   Total \tab 3595 \tab 240 \tab 179 \tab 281 \tab 200 \tab 1 \tab 14 \tab 4510 \cr
}
}

\seealso{
Other Mental Health III: 
\code{\link{mhothrel}},
\code{\link{mhtrtoth}},
\code{\link{seemhpub}},
\code{\link{viglab06}}
}
\concept{Mental Health III}
\keyword{variable}
