% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_414.R
\name{mhsecret}
\alias{mhsecret}
\title{X s family better off if kept secret}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhsecret
}
\details{
Question For the next several questions, please tell me whether you strongly agree, agree, disagree, or strongly disagree with the statement.
R. Members of (NAME)'s family would be better off if (NAME)'s situation was kept secret.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3779/vshow}{GSS Data Explorer page for \code{mhsecret}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 10 \tab 131 \tab 898 \tab 350 \tab 27 \tab 9 \tab 4510 \cr
   Total \tab 3085 \tab 10 \tab 131 \tab 898 \tab 350 \tab 27 \tab 9 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhgvtjob}},
\code{\link{mhhlpfam}},
\code{\link{mhhlpfrd}},
\code{\link{mhhlpgp}},
\code{\link{mhhlpmhp}},
\code{\link{mhhlppsy}},
\code{\link{mhhlprel}},
\code{\link{mhnotch}},
\code{\link{mhsupwrk}}
}
\concept{Mental Health IV}
\keyword{variable}
