% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_416.R
\name{mhseroth}
\alias{mhseroth}
\title{How serious his her mental health problem was}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhseroth
}
\details{
Question Let's turn away from (NAME) now. I have a few questions about your general experiences that I would like to ask you. Leaving yourself aside, have you personally ever known someone who has received treatment for a mental health situation?
B. How serious would you say (his/her) mental health problem was, was it…
}
\section{Values}{

\itemize{
\item \code{1} very serious
\item \code{2} somewhat serious
\item \code{3} not very serious
\item \code{4} not at all serious
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3804/vshow}{GSS Data Explorer page for \code{mhseroth}}.
}

\seealso{
Other Mental Health IV: 
\code{\link{mhcause}},
\code{\link{mhclsoth}},
\code{\link{mhdisoth}},
\code{\link{mhhlpoth}},
\code{\link{mhresoth}},
\code{\link{mhtrtot2}},
\code{\link{mhviooth}},
\code{\link{mhvioslf}}

Other Mental Health: 
\code{\link{mhcause}},
\code{\link{mhclsoth}},
\code{\link{mhhlpoth}},
\code{\link{mhresoth}},
\code{\link{mhtrtot2}},
\code{\link{mhviooth}},
\code{\link{mhvioslf}}
}
\concept{Mental Health}
\concept{Mental Health IV}
\keyword{variable}
