% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_392.R
\name{mhtrtoth}
\alias{mhtrtoth}
\title{R personally has known someone who received treatment}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhtrtoth
}
\details{
Question Leaving yourself aside, have you personally ever known someone who has received treatment for a mental health situation?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3565/vshow}{GSS Data Explorer page for \code{mhtrtoth}}.
}

\seealso{
Other Mental Health III: 
\code{\link{avoidmh}},
\code{\link{gesttalk}},
\code{\link{mhothrel}},
\code{\link{mhothyou}},
\code{\link{seemhpub}},
\code{\link{viglab06}}
}
\concept{Mental Health III}
\keyword{variable}
