% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_417.R
\name{mhtrtslf}
\alias{mhtrtslf}
\title{R have ever personally received treatment for mental health problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhtrtslf
}
\details{
Question Have you personally ever received treatment for a mental health problem?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3810/vshow}{GSS Data Explorer page for \code{mhtrtslf}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 213 \tab 1200 \tab 1 \tab 11 \tab 4510 \cr
   Total \tab 3085 \tab 213 \tab 1200 \tab 1 \tab 11 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhexpoth}},
\code{\link{mhfright}},
\code{\link{mhreloth}},
\code{\link{mhseepub}},
\code{\link{mhsymp}}
}
\concept{Mental Health IV}
\keyword{variable}
