% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_413.R
\name{mhtrusty}
\alias{mhtrusty}
\title{People like x are just as trustworthy as anyone else}
\source{
General Social Survey https://gss.norc.org
}
\description{
mhtrusty
}
\details{
Question For the next several questions, please tell me whether you strongly agree, agree, disagree, or strongly disagree with the statement.
N. People like (NAME) are just as trustworthy as anyone else.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3775/vshow}{GSS Data Explorer page for \code{mhtrusty}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 217 \tab 816 \tab 285 \tab 28 \tab 70 \tab 9 \tab 4510 \cr
   Total \tab 3085 \tab 217 \tab 816 \tab 285 \tab 28 \tab 70 \tab 9 \tab 4510 \cr
}
}

\seealso{
Other Mental Health IV: 
\code{\link{mhaccptd}},
\code{\link{mhcreatv}},
\code{\link{mhfeelem}},
\code{\link{mhhired}},
\code{\link{mhhrdtlk}},
\code{\link{mhnervs}},
\code{\link{mhnokids}},
\code{\link{mhnotell}},
\code{\link{mhprdctv}}
}
\concept{Mental Health IV}
\keyword{variable}
