% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_90.R
\name{milokfe}
\alias{milokfe}
\title{Mil service good experience for women}
\source{
General Social Survey https://gss.norc.org
}
\description{
milokfe
}
\details{
Question B. How about for most young women, do you think military service is definitely a good experience, probably a good experience, probably not a good experience, or definitely not a good experience for them?
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely good
\item \verb{[2]} probably good
\item \verb{[3]} prob not good
\item \verb{[4]} def not good
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/819/vshow}{GSS Data Explorer page for \code{milokfe}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely good \tab probably good \tab prob not good \tab def not good \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 208 \tab 807 \tab 283 \tab 95 \tab 58 \tab 22 \tab 1473 \cr
   Total \tab 0 \tab 208 \tab 807 \tab 283 \tab 95 \tab 58 \tab 22 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{cojail}},
\code{\link{copunish}},
\code{\link{draftco}},
\code{\link{draftdef}},
\code{\link{femilop}},
\code{\link{milokme}},
\code{\link{minmilop}},
\code{\link{vetfam}},
\code{\link{vetfamnw}}
}
\concept{Core}
\keyword{variable}
