% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_87.R
\name{milqual}
\alias{milqual}
\title{Quality of persons in military}
\source{
General Social Survey https://gss.norc.org
}
\description{
milqual
}
\details{
Question As you know, this country stopped the military draft in 1972. Since that time we have relied on volunteers. Now I'd like to ask you a few questions about our armed forces.
A. How would you rate the quality of the men and women now serving in the armed forces?  Would you say the quality of personnel is excellent, good, not so good, or poor? (Just your own opinion based on what you've heard or read.)
}
\section{Values}{

\itemize{
\item \code{1} excellent
\item \code{2} good
\item \code{3} not so good
\item \code{4} poor
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
   1983 \tab - \tab full \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/781/vshow}{GSS Data Explorer page for \code{milqual}}.
}

\seealso{
Other Core: 
\code{\link{blnumok}},
\code{\link{blnumoky}},
\code{\link{fenumok}},
\code{\link{fightair}},
\code{\link{hinumok}},
\code{\link{hinumoky}},
\code{\link{milpay}},
\code{\link{milvolok}},
\code{\link{taxserve}}

Other Military: 
\code{\link{blnumok}},
\code{\link{blnumoky}},
\code{\link{fenumok}},
\code{\link{fightair}},
\code{\link{hinumok}},
\code{\link{hinumoky}},
\code{\link{milpay}},
\code{\link{milvolok}},
\code{\link{taxserve}}
}
\concept{Core}
\concept{Military}
\keyword{variable}
