% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_538.R
\name{milserve}
\alias{milserve}
\title{How important to serve in military when needed}
\source{
General Social Survey https://gss.norc.org
}
\description{
milserve
}
\details{
Question There are different opinions as to what it takes to be a good citizen. As far as you are concerned personally on a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it…
J. To be willing to serve in the military at a time of need
}
\section{Values}{

\itemize{
\item \code{1} not at all important
\item \code{7} very important
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4916/vshow}{GSS Data Explorer page for \code{milserve}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab not at all important \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 112 \tab 51 \tab 61 \tab 130 \tab 232 \tab 226 \tab 641 \tab 15 \tab 4 \tab 2812 \cr
   Total \tab 1340 \tab 112 \tab 51 \tab 61 \tab 130 \tab 232 \tab 226 \tab 641 \tab 15 \tab 4 \tab 2812 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actassoc}},
\code{\link{buypol}},
\code{\link{helpusa}},
\code{\link{helpwrld}},
\code{\link{obeylaws}},
\code{\link{othreasn}},
\code{\link{paytaxes}},
\code{\link{relmeet}},
\code{\link{watchgov}}
}
\concept{ISSP Citizenship}
\keyword{variable}
