% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_87.R
\name{milvolok}
\alias{milvolok}
\title{How well has volunteer military worked?}
\source{
General Social Survey https://gss.norc.org
}
\description{
milvolok
}
\details{
Question All things considered, how well do you think relying on volunteers has worked for the armed forces--has it worked
very well, fairly well, or not well?
}
\section{Values}{

\itemize{
\item \code{1} very well
\item \code{2} fairly well
\item \code{3} not well
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/788/vshow}{GSS Data Explorer page for \code{milvolok}}.
}

\seealso{
Other Core: 
\code{\link{blnumok}},
\code{\link{blnumoky}},
\code{\link{fenumok}},
\code{\link{fightair}},
\code{\link{hinumok}},
\code{\link{hinumoky}},
\code{\link{milpay}},
\code{\link{milqual}},
\code{\link{taxserve}}
}
\concept{Core}
\keyword{variable}
