% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_92.R
\name{milwrkev}
\alias{milwrkev}
\title{Ever work for military or dod?}
\source{
General Social Survey https://gss.norc.org
}
\description{
milwrkev
}
\details{
Question Have you ever had a civilian job with the armed forces or the Defense Department?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/831/vshow}{GSS Data Explorer page for \code{milwrkev}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1984 \tab - \tab 82 \tab 1363 \tab 28 \tab 1473 \cr
   Total \tab 0 \tab 82 \tab 1363 \tab 28 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{defwrknw}},
\code{\link{milwrknw}},
\code{\link{ob911}},
\code{\link{obeng}},
\code{\link{objury}},
\code{\link{obknow}},
\code{\link{obvol}},
\code{\link{obvote}},
\code{\link{resdefwk}}
}
\concept{Core}
\keyword{variable}
