% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_532.R
\name{mincult}
\alias{mincult}
\title{Ethnic minorities should be given gov assistance}
\source{
General Social Survey https://gss.norc.org
}
\description{
mincult
}
\details{
Question How much do you agree or disagree with the following statements?
B. Ethnic minorities should be given government assistance to preserve their customs and traditions
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} disagree strongly
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4857/vshow}{GSS Data Explorer page for \code{mincult}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 46 \tab 168 \tab 293 \tab 502 \tab 242 \tab 55 \tab 61 \tab 2904 \cr
   2004 \tab 1596 \tab 50 \tab 222 \tab 300 \tab 455 \tab 160 \tab 25 \tab 4 \tab 2812 \cr
   2014 \tab 1264 \tab 13 \tab 189 \tab 245 \tab 537 \tab 235 \tab 52 \tab 3 \tab 2538 \cr
   Total \tab 4397 \tab 109 \tab 579 \tab 838 \tab 1494 \tab 637 \tab 132 \tab 68 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amcult}},
\code{\link{amownway}},
\code{\link{amtv}},
\code{\link{forland}},
\code{\link{forlang}},
\code{\link{imports}},
\code{\link{proudgrp}},
\code{\link{proudhis}},
\code{\link{wrldgovt}}
}
\concept{ISSP National Identity}
\keyword{variable}
