% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_90.R
\name{minmilop}
\alias{minmilop}
\title{Min opportunities better in military?}
\source{
General Social Survey https://gss.norc.org
}
\description{
minmilop
}
\details{
Question A. Thinking about opportunities and equal treatment for minority groups, like Blacks and Hispanics - Would you say that their treatment and opportunities are better in the military, better in civilian employment, or that there isn't any difference these days?
}
\section{Values}{

\itemize{
\item \code{1} military better
\item \code{2} civilian better
\item \code{3} no difference
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/814/vshow}{GSS Data Explorer page for \code{minmilop}}.
}

\seealso{
Other Core: 
\code{\link{cojail}},
\code{\link{copunish}},
\code{\link{draftco}},
\code{\link{draftdef}},
\code{\link{femilop}},
\code{\link{milokfe}},
\code{\link{milokme}},
\code{\link{vetfam}},
\code{\link{vetfamnw}}
}
\concept{Core}
\keyword{variable}
