% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_568.R
\name{minthree}
\alias{minthree}
\title{Imputed poverty line for family of three}
\source{
General Social Survey https://gss.norc.org
}
\description{
minthree
}
\details{
Question And finally, what amount of weekly income would you use as a "poverty line" for a family of three, made up of a woman and two children, in this community?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5204/vshow}{GSS Data Explorer page for \code{minthree}}.
}

\seealso{
Other Income: 
\code{\link{minfood}}

Other Income: 
\code{\link{minfood}}
}
\concept{Income}
\keyword{variable}
