% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_511.R
\name{miracles}
\alias{miracles}
\title{Belief in religious miracles}
\source{
General Social Survey https://gss.norc.org
}
\description{
miracles
}
\details{
Question Do you believe in. . .
E. Religious miracles
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, definitely
\item \verb{[2]} yes, probably
\item \verb{[3]} no, probably not
\item \verb{[4]} no, definitely not
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4655/vshow}{GSS Data Explorer page for \code{miracles}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes, definitely \tab yes, probably \tab no, probably not \tab no, definitely not \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 543 \tab 331 \tab 193 \tab 123 \tab 95 \tab 74 \tab 1517 \cr
   1998 \tab 1548 \tab 577 \tab 304 \tab 142 \tab 99 \tab 102 \tab 60 \tab 2832 \cr
   2008 \tab 658 \tab 738 \tab 311 \tab 183 \tab 101 \tab 29 \tab 3 \tab 2023 \cr
   2018 \tab 1173 \tab 591 \tab 277 \tab 159 \tab 114 \tab 25 \tab 9 \tab 2348 \cr
   Total \tab 3537 \tab 2449 \tab 1223 \tab 677 \tab 437 \tab 251 \tab 146 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{bible1}},
\code{\link{devil}},
\code{\link{fatalism}},
\code{\link{godmeans}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{nihilism}},
\code{\link{predeter}},
\code{\link{theism}}
}
\concept{ISSP Religion}
\keyword{variable}
