% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_219.R
\name{mnilwhy1}
\alias{mnilwhy1}
\title{1st mention cause of mental illness feelings}
\source{
General Social Survey https://gss.norc.org
}
\description{
mnilwhy1
}
\details{
Question (IF NO TO EVBRKDWN, ASK EVMHP) Have you ever felt you had a mental health problem?
A. (IF YES TO EVMHP ASK:) Could you tell me about when you felt that way?  What was it about? (PROBE: GET EXACT DATE and complete details.)
A1. The causes and sources of R's mental illness -- First mention.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2079/vshow}{GSS Data Explorer page for \code{mnilwhy1}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 12 \tab 123 \tab 15 \tab 19 \tab 200 \tab 204 \tab 21 \tab 217 \tab 219 \tab 22 \tab 309 \tab 31 \tab 35 \tab 38 \tab 39 \tab 41 \tab 500 \tab 502 \tab 524 \tab 700 \tab 701 \tab 702 \tab 709 \tab 80 \tab 802 \tab 803 \tab 804 \tab 806 \tab 807 \tab 809 \tab no answer \tab Total \cr
   1996 \tab 2815 \tab 4 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 8 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 15 \tab 1 \tab 1 \tab 2 \tab 1 \tab 19 \tab 2904 \cr
   Total \tab 2815 \tab 4 \tab 1 \tab 3 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 8 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 15 \tab 1 \tab 1 \tab 2 \tab 1 \tab 19 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkddo2}},
\code{\link{brkddo3}},
\code{\link{brkddur}},
\code{\link{brkdhlp1}},
\code{\link{brkdhlp2}},
\code{\link{brkdhlp3}},
\code{\link{brkdtime}},
\code{\link{evmhp}},
\code{\link{mnilwhy2}}
}
\concept{Mental Health}
\keyword{variable}
