% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_219.R
\name{mnilwhy2}
\alias{mnilwhy2}
\title{2nd mention cause of mental illness feelings}
\source{
General Social Survey https://gss.norc.org
}
\description{
mnilwhy2
}
\details{
Question (IF NO TO EVBRKDWN, ASK EVMHP) Have you ever felt you had a mental health problem?
A2. The causes and sources of R''s mental illness -- Second mention.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2080/vshow}{GSS Data Explorer page for \code{mnilwhy2}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab iap \tab 129 \tab 15 \tab 16 \tab 17 \tab 204 \tab 21 \tab 216 \tab 23 \tab 35 \tab 41 \tab 649 \tab 800 \tab 802 \tab 803 \tab 805 \tab 809 \tab no answer \tab Total \cr
   1996 \tab 2860 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 19 \tab 2904 \cr
   Total \tab 2860 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 19 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkddo2}},
\code{\link{brkddo3}},
\code{\link{brkddur}},
\code{\link{brkdhlp1}},
\code{\link{brkdhlp2}},
\code{\link{brkdhlp3}},
\code{\link{brkdtime}},
\code{\link{evmhp}},
\code{\link{mnilwhy1}}
}
\concept{Mental Health}
\keyword{variable}
