% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_577.R
\name{mnthsusa}
\alias{mnthsusa}
\title{Lived in u.s., if less than one year 		months:}
\source{
General Social Survey https://gss.norc.org
}
\description{
mnthsusa
}
\details{
Question How long have you lived in the United States? If you have lived in the United States for more than one period during your life, what is the total number of years you have lived in the United States? - Months
}
\section{Values}{

\itemize{
\item \code{1} 1 month
\item \code{11} 11 months
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7298/vshow}{GSS Data Explorer page for \code{mnthsusa}}.
}

\keyword{variable}
