% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_215.R
\name{mntlas1}
\alias{mntlas1}
\title{1st mention, what mental illness is, asserted}
\source{
General Social Survey https://gss.norc.org
}
\description{
mntlas1
}
\details{
Question Of course, everybody hears a good deal about physical illness and disease, but now, what about the ones we call mental or nervous illness?...When you hear someone say that a person is "mentally-ill," what does that mean to you? (PROBES: How would you describe a person who is mentally-ill?  What do you think a mentally-ill person is like?  What does a person like this do that tells you he is mentally-ill?  How does a person like this act?)
\enumerate{
\item R's asserted references for what mental illness is -- First mention.
}
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2040/vshow}{GSS Data Explorer page for \code{mntlas1}}.
}

\seealso{
Other Mental Health: 
\code{\link{dangroth}},
\code{\link{dangrslf}},
\code{\link{govmentl}},
\code{\link{medcare1}},
\code{\link{medcare2}},
\code{\link{mustdoc}},
\code{\link{musthosp}},
\code{\link{mustmed}},
\code{\link{spmentl}}
}
\concept{Mental Health}
\keyword{variable}
