% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_243.R
\name{mntlcare}
\alias{mntlcare}
\title{Were able to get mental health care}
\source{
General Social Survey https://gss.norc.org
}
\description{
mntlcare
}
\details{
Question Were you able to get the treatment you sought?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2190/vshow}{GSS Data Explorer page for \code{mntlcare}}.
}

\seealso{
Other Mental Health: 
\code{\link{anydoc}},
\code{\link{doclist}},
\code{\link{hlthinsr}},
\code{\link{psycdrct}},
\code{\link{psycfdoc}},
\code{\link{psyckid3}},
\code{\link{psycpsyc}},
\code{\link{swithlth}},
\code{\link{usedmntl}}

Other Health: 
\code{\link{anydoc}},
\code{\link{doclist}},
\code{\link{hlthinsr}},
\code{\link{swithlth}},
\code{\link{usedmntl}}

Other Mental Health: 
\code{\link{anydoc}},
\code{\link{doclist}},
\code{\link{hlthinsr}},
\code{\link{psycdrct}},
\code{\link{psycfdoc}},
\code{\link{psyckid3}},
\code{\link{psycpsyc}},
\code{\link{swithlth}},
\code{\link{usedmntl}}
}
\concept{Health}
\concept{Mental Health}
\keyword{variable}
