% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_244.R
\name{mntldeny}
\alias{mntldeny}
\title{Ever denied services under plans package}
\source{
General Social Survey https://gss.norc.org
}
\description{
mntldeny
}
\details{
Question Were you denied services under your plan's benefit package?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2192/vshow}{GSS Data Explorer page for \code{mntldeny}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1998 \tab 2678 \tab 16 \tab 130 \tab 1 \tab 7 \tab 2832 \cr
   Total \tab 2678 \tab 16 \tab 130 \tab 1 \tab 7 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{doc1}},
\code{\link{hmo1}},
\code{\link{hmo2}},
\code{\link{hmo3}},
\code{\link{hmo4}},
\code{\link{hmo5}},
\code{\link{hmo6}},
\code{\link{hmo7}},
\code{\link{mntldif}}
}
\concept{Mental Health}
\keyword{variable}
