% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_244.R
\name{mntldif}
\alias{mntldif}
\title{How difficult to get mental health care}
\source{
General Social Survey https://gss.norc.org
}
\description{
mntldif
}
\details{
Question How difficult was it to get the treatment that you sought: very difficult, somewhat difficult, or not at all difficult?
}
\section{Values}{

\itemize{
\item \code{1} very difficult
\item \code{2} somewhat diffclt
\item \code{3} not difficult
\item \code{8} know know
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2191/vshow}{GSS Data Explorer page for \code{mntldif}}.
}

\seealso{
Other Mental Health: 
\code{\link{doc1}},
\code{\link{hmo1}},
\code{\link{hmo2}},
\code{\link{hmo3}},
\code{\link{hmo4}},
\code{\link{hmo5}},
\code{\link{hmo6}},
\code{\link{hmo7}},
\code{\link{mntldeny}}
}
\concept{Mental Health}
\keyword{variable}
