% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_332.R
\name{mntlrel}
\alias{mntlrel}
\title{Effect of mental health problem on relation with}
\source{
General Social Survey https://gss.norc.org
}
\description{
mntlrel
}
\details{
Question I want you to think about the person with a mental health problem with whom you have had the most contact. As a result of the mental health problem, would you say that your relationship with this person...
}
\section{Values}{

\itemize{
\item \code{1} became stronger
\item \code{2} was unchanged
\item \code{3} became worse
\item \code{4} ended as a result of the problem
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2973/vshow}{GSS Data Explorer page for \code{mntlrel}}.
}

\seealso{
Other Mental Health II: 
\code{\link{adcou}},
\code{\link{adcoumed}},
\code{\link{addoc}},
\code{\link{addocmed}},
\code{\link{adfam}},
\code{\link{adfammed}},
\code{\link{adhdcon}},
\code{\link{adhdknow}},
\code{\link{adhdknw}},
\code{\link{adhdmed}},
\code{\link{adhdreal}},
\code{\link{adhos}},
\code{\link{adhosmed}},
\code{\link{adpsy}},
\code{\link{adpsymed}},
\code{\link{adtch}},
\code{\link{adtchmed}},
\code{\link{allergic}},
\code{\link{badchar}},
\code{\link{bettrown}},
\code{\link{chembal}},
\code{\link{chldfrnd}},
\code{\link{chldhlp}},
\code{\link{chldprb}},
\code{\link{chldprob}},
\code{\link{chldsch}},
\code{\link{chldvig}},
\code{\link{comknows}},
\code{\link{discipln}},
\code{\link{doc15a}},
\code{\link{doc16a}},
\code{\link{doc18a}},
\code{\link{doc19a}},
\code{\link{doc20a}},
\code{\link{failure}},
\code{\link{forcedoc}},
\code{\link{forcehos}},
\code{\link{forcemed}},
\code{\link{geneprob}},
\code{\link{grwingup}},
\code{\link{illphys}},
\code{\link{imprvdis}},
\code{\link{imprveat}},
\code{\link{imprvmed}},
\code{\link{medaddct}},
\code{\link{medkdneg}},
\code{\link{medsavtx}},
\code{\link{medsymps}},
\code{\link{medunacc}},
\code{\link{medweak}},
\code{\link{mntlill}},
\code{\link{nextdoor}},
\code{\link{outsider}},
\code{\link{ovrmedkd}},
\code{\link{pryntfam}},
\code{\link{putsoff}},
\code{\link{raised}},
\code{\link{solveown}},
\code{\link{spendeve}},
\code{\link{stressfl}},
\code{\link{sufadult}},
\code{\link{trbllaw}},
\code{\link{violpeop}},
\code{\link{violself}},
\code{\link{violtv}},
\code{\link{zombies}}
}
\concept{Mental Health II}
\keyword{variable}
