% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_332.R
\name{mntlrel}
\alias{mntlrel}
\title{Effect of mental health problem on relation with}
\source{
General Social Survey https://gss.norc.org
}
\description{
mntlrel
}
\details{
Question I want you to think about the person with a mental health problem with whom you have had the most contact. As a result of the mental health problem, would you say that your relationship with this person...
}
\section{Values}{

\itemize{
\item \code{1} became stronger
\item \code{2} was unchanged
\item \code{3} became worse
\item \code{4} ended as a result of the problem
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2973/vshow}{GSS Data Explorer page for \code{mntlrel}}.
}

\seealso{
Other Mental Health II: 
\code{\link{adhdcon}},
\code{\link{adhdknow}},
\code{\link{adhdknw}},
\code{\link{adhdmed}},
\code{\link{adhdreal}},
\code{\link{medsavtx}},
\code{\link{pryntfam}},
\code{\link{trbllaw}},
\code{\link{zombies}}
}
\concept{Mental Health II}
\keyword{variable}
