% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_216.R
\name{mntlsym3}
\alias{mntlsym3}
\title{3rd mention, what mental illness is, symptoms}
\source{
General Social Survey https://gss.norc.org
}
\description{
mntlsym3
}
\details{
Question Of course, everybody hears a good deal about physical illness and disease, but now, what about the ones we call mental or nervous illness?...When you hear someone say that a person is "mentally-ill," what does that mean to you? (PROBES: How would you describe a person who is mentally-ill?  What do you think a mentally-ill person is like?  What does a person like this do that tells you he is mentally-ill?  How does a person like this act?)
3. Symptoms or manifestations of mental illness -- Third mention.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2047/vshow}{GSS Data Explorer page for \code{mntlsym3}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 202 \tab 203 \tab 204 \tab 205 \tab 206 \tab 207 \tab 208 \tab 210 \tab 211 \tab 212 \tab 303 \tab 304 \tab 306 \tab 307 \tab 308 \tab 310 \tab 402 \tab 403 \tab 404 \tab 405 \tab 408 \tab 411 \tab 412 \tab 501 \tab 502 \tab 503 \tab 510 \tab 511 \tab 512 \tab 602 \tab 701 \tab 704 \tab 705 \tab 706 \tab 707 \tab 708 \tab 709 \tab 710 \tab 712 \tab 803 \tab 807 \tab 808 \tab 811 \tab 812 \tab 901 \tab 902 \tab 904 \tab 905 \tab 906 \tab 907 \tab 908 \tab 910 \tab 911 \tab 921 \tab 922 \tab 923 \tab 925 \tab no answer \tab Total \cr
   1996 \tab 2572 \tab 6 \tab 3 \tab 2 \tab 8 \tab 1 \tab 14 \tab 8 \tab 8 \tab 3 \tab 11 \tab 13 \tab 16 \tab 11 \tab 2 \tab 2 \tab 1 \tab 5 \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab 3 \tab 3 \tab 2 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 3 \tab 2 \tab 1 \tab 2 \tab 2 \tab 4 \tab 3 \tab 9 \tab 2 \tab 1 \tab 2 \tab 6 \tab 10 \tab 5 \tab 1 \tab 6 \tab 5 \tab 4 \tab 11 \tab 1 \tab 7 \tab 1 \tab 1 \tab 3 \tab 7 \tab 4 \tab 3 \tab 86 \tab 2904 \cr
   Total \tab 2572 \tab 6 \tab 3 \tab 2 \tab 8 \tab 1 \tab 14 \tab 8 \tab 8 \tab 3 \tab 11 \tab 13 \tab 16 \tab 11 \tab 2 \tab 2 \tab 1 \tab 5 \tab 1 \tab 3 \tab 3 \tab 1 \tab 1 \tab 3 \tab 3 \tab 2 \tab 2 \tab 1 \tab 2 \tab 3 \tab 1 \tab 3 \tab 2 \tab 1 \tab 2 \tab 2 \tab 4 \tab 3 \tab 9 \tab 2 \tab 1 \tab 2 \tab 6 \tab 10 \tab 5 \tab 1 \tab 6 \tab 5 \tab 4 \tab 11 \tab 1 \tab 7 \tab 1 \tab 1 \tab 3 \tab 7 \tab 4 \tab 3 \tab 86 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkdas1}},
\code{\link{brkdas2}},
\code{\link{mntlas2}},
\code{\link{mntlimp1}},
\code{\link{mntlimp2}},
\code{\link{mntlimp3}},
\code{\link{mntloth}},
\code{\link{mntlsym1}},
\code{\link{mntlsym2}}
}
\concept{Mental Health}
\keyword{variable}
