% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_607.R
\name{mode}
\alias{mode}
\title{Interview done in-person or over the phone}
\source{
General Social Survey https://gss.norc.org
}
\description{
mode
}
\details{
Question Interview done in-person or over the phone
}
\section{Values}{

\itemize{
\item \code{1} in-person
\item \code{2} by phone
\item \code{3} multimode
\item \code{4} web
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5566/vshow}{GSS Data Explorer page for \code{mode}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab in-person \tab by phone \tab don't know \tab no answer \tab web \tab combination in-person/phone \tab Total \cr
   2004 \tab - \tab 2219 \tab 593 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 4209 \tab 301 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 1778 \tab 245 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 1824 \tab 211 \tab 1 \tab 8 \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 1712 \tab 258 \tab - \tab 4 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 2199 \tab 333 \tab - \tab 6 \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 2665 \tab 195 \tab - \tab 7 \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 2094 \tab 245 \tab - \tab 9 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab 471 \tab - \tab - \tab 3561 \tab - \tab 4032 \cr
   2022 \tab - \tab 1573 \tab 206 \tab - \tab - \tab 1636 \tab 129 \tab 3544 \cr
   Total \tab 0 \tab 20273 \tab 3058 \tab 1 \tab 34 \tab 5197 \tab 129 \tab 28692 \cr
}
}

\seealso{
Other Paradata: 
\code{\link{consent}},
\code{\link{easyget}},
\code{\link{intage}},
\code{\link{intethn}},
\code{\link{intsex}},
\code{\link{intyrs}},
\code{\link{lngthinv}}
}
\concept{Paradata}
\keyword{variable}
