% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_607.R
\name{mode}
\alias{mode}
\title{Interview done in-person or over the phone}
\source{
General Social Survey https://gss.norc.org
}
\description{
mode
}
\details{
Question Interview done in-person or over the phone
}
\section{Values}{

\itemize{
\item \code{1} in-person
\item \code{2} by phone
\item \code{3} multimode
\item \code{4} web
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
   2024 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5566/vshow}{GSS Data Explorer page for \code{mode}}.
}

\seealso{
Other Paradata: 
\code{\link{consent}},
\code{\link{easyget}},
\code{\link{intage}},
\code{\link{intethn}},
\code{\link{intsex}},
\code{\link{intyrs}},
\code{\link{lngthinv}}
}
\concept{Paradata}
\keyword{variable}
