% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_677.R
\name{modelobby}
\alias{modelobby}
\title{Has r contacted a politician in last 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
modelobby
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} only offline
\item \code{2} only online
\item \code{3} both offline and online
\item \code{4} neither offline nor online
\item \code{NA(d)} can't choose
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
