% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_677.R
\name{modeorg}
\alias{modeorg}
\title{Has r belonged to an activist group in last 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
modeorg
}
\details{
Question Here are some different forms of political and social action that people can take. Please indicate, for each one, whether you have done any of these things in the last 12 months offline only, online only, or both offline and online.… Belonged to or joined a group that is involved in social issues or working to advance a cause.
}
\section{Values}{

\itemize{
\item \code{1} only offline
\item \code{2} only online
\item \code{3} both offline and online
\item \code{4} neither offline nor online
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8154/vshow}{GSS Data Explorer page for \code{modeorg}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{modelobby}},
\code{\link{modeorgprot}},
\code{\link{modepet}},
\code{\link{modeprot}},
\code{\link{papernews}},
\code{\link{radionews}},
\code{\link{trppl}},
\code{\link{tvnews1}},
\code{\link{webnews}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
