% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_677.R
\name{modeorgprot}
\alias{modeorgprot}
\title{Has r organized a protest in last 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
modeorgprot
}
\details{
Question Here are some different forms of political and social action that people can take. Please indicate, for each one, whether you have done any of these things in the last 12 months offline only, online only, or both offline and online… Organized or helped to organize a demonstration or political protest.
}
\section{Values}{

\itemize{
\item \code{1} only offline
\item \code{2} only online
\item \code{3} both offline and online
\item \code{4} neither offline nor online
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8155/vshow}{GSS Data Explorer page for \code{modeorgprot}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{modelobby}},
\code{\link{modeorg}},
\code{\link{modepet}},
\code{\link{modeprot}},
\code{\link{papernews}},
\code{\link{radionews}},
\code{\link{trppl}},
\code{\link{tvnews1}},
\code{\link{webnews}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
