% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_665.R
\name{modesequence}
\alias{modesequence}
\title{Initial mode of contact for respondent}
\source{
General Social Survey https://gss.norc.org
}
\description{
modesequence
}
\details{
Question SEQUENCE OF MODE ASSIGNMENT FOR RESPONDENT
}
\section{Values}{

\itemize{
\item \verb{[1]} in-person first
\item \verb{[2]} web-first
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7970/vshow}{GSS Data Explorer page for \code{modesequence}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab in-person first \tab web-first \tab Total \cr
   2022 \tab - \tab 1540 \tab 2004 \tab 3544 \cr
   Total \tab 0 \tab 1540 \tab 2004 \tab 3544 \cr
}
}

\keyword{variable}
