% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_151.R
\name{modpaint}
\alias{modpaint}
\title{Modern painting is just slapped on}
\source{
General Social Survey https://gss.norc.org
}
\description{
modpaint
}
\details{
Question I'm going to read several statements. After I read each one, please tell me whether you strongly agree, agree, disagree, or strongly disagree with it. For example, ...
E. Modern painting is just slapped on: a child could do it.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1426/vshow}{GSS Data Explorer page for \code{modpaint}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 72 \tab 528 \tab 720 \tab 153 \tab 117 \tab 16 \tab 1606 \cr
   1998 \tab 1387 \tab 60 \tab 335 \tab 695 \tab 184 \tab 154 \tab 17 \tab 2832 \cr
   Total \tab 1387 \tab 132 \tab 863 \tab 1415 \tab 337 \tab 271 \tab 33 \tab 4438 \cr
}
}

\seealso{
Other Culture: 
\code{\link{classics}},
\code{\link{english}},
\code{\link{grtbooks}},
\code{\link{judgeart}},
\code{\link{pclit}},
\code{\link{trstprof}},
\code{\link{tvnews}},
\code{\link{tvpbs}},
\code{\link{tvshows}}
}
\concept{Culture}
\keyword{variable}
