% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_202.R
\name{monfound}
\alias{monfound}
\title{Done volunteer work in priv & comm. found}
\source{
General Social Survey https://gss.norc.org
}
\description{
monfound
}
\details{
Question In which, if any, of the areas listed on this card have you done some volunteer work in the past month? (RECORD above UNDER MONHLTH-MONOTH. CIRCLE "YES" OR "NO" FOR EACH AREA)
L. Private and community foundations
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1911/vshow}{GSS Data Explorer page for \code{monfound}}.
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{monart}},
\code{\link{monenvir}},
\code{\link{monhuman}},
\code{\link{monpol}},
\code{\link{monpub}},
\code{\link{monrec}},
\code{\link{monrelig}},
\code{\link{monwork}},
\code{\link{monyouth}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
