% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_203.R
\name{monintl}
\alias{monintl}
\title{Done volunteer work in international}
\source{
General Social Survey https://gss.norc.org
}
\description{
monintl
}
\details{
Question In which, if any, of the areas listed on this card have you done some volunteer work in the past month? (RECORD above UNDER MONHLTH-MONOTH. CIRCLE "YES" OR "NO" FOR EACH AREA)
M. International/foreign
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1912/vshow}{GSS Data Explorer page for \code{monintl}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1996 \tab 2875 \tab 8 \tab 19 \tab 2 \tab 2904 \cr
   Total \tab 2875 \tab 8 \tab 19 \tab 2 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{hrseduc}},
\code{\link{hrsenvir}},
\code{\link{hrshlth}},
\code{\link{hrshuman}},
\code{\link{hrspub}},
\code{\link{hrsrec}},
\code{\link{hrsrelig}},
\code{\link{moninfrm}},
\code{\link{monoth}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
