% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_140.R
\name{monitred}
\alias{monitred}
\title{Computers monitor my job performance}
\source{
General Social Survey https://gss.norc.org
}
\description{
monitred
}
\details{
Question How true would you say each of the following statements is about your job?  Would you say very true, true, somewhat true, or not true at all:
E. Computers monitor my job performance.
}
\section{Values}{

\itemize{
\item \code{1} very true
\item \code{2} true
\item \code{3} somewhat true
\item \code{4} not true at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1312/vshow}{GSS Data Explorer page for \code{monitred}}.
}

\seealso{
Other Work Organizations: 
\code{\link{automatn}},
\code{\link{idecide}},
\code{\link{levels}()},
\code{\link{offsup}},
\code{\link{supduty1}},
\code{\link{supothrs}},
\code{\link{totsup}},
\code{\link{yousup}}
}
\concept{Work Organizations}
\keyword{variable}
