% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_241.R
\name{monthly}
\alias{monthly}
\title{What is your monthly wage before tax}
\source{
General Social Survey https://gss.norc.org
}
\description{
monthly
}
\details{
Question IF HOURLY WAGE:
D. What is your monthly salary, before any taxes or deductions?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2164/vshow}{GSS Data Explorer page for \code{monthly}}.
}

\seealso{
Other Jobs: 
\code{\link{daily}},
\code{\link{doingdys}},
\code{\link{doingfin}},
\code{\link{doingtrn}},
\code{\link{formldys}},
\code{\link{formlfin}},
\code{\link{formlhrs}},
\code{\link{formltrn}},
\code{\link{genderwk}},
\code{\link{hourly}},
\code{\link{howpaid}},
\code{\link{locyrs}},
\code{\link{numemp}},
\code{\link{occtrain}},
\code{\link{occyrs}},
\code{\link{orgsize}},
\code{\link{othyrs}},
\code{\link{otjtrain}},
\code{\link{partorg}},
\code{\link{samework}},
\code{\link{siteyrs}},
\code{\link{trainsch}},
\code{\link{trusting}},
\code{\link{typeorg}},
\code{\link{weekly}},
\code{\link{wojob}},
\code{\link{wojobyrs}},
\code{\link{wrkoth}},
\code{\link{yearly}}
}
\concept{Jobs}
\keyword{variable}
