% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_202.R
\name{monyouth}
\alias{monyouth}
\title{H done volunteer work in youth development}
\source{
General Social Survey https://gss.norc.org
}
\description{
monyouth
}
\details{
Question In which, if any, of the areas listed on this card have you done some volunteer work in the past month? (RECORD above UNDER MONHLTH-MONOTH. CIRCLE "YES" OR "NO" FOR EACH AREA)
K. Youth development
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1910/vshow}{GSS Data Explorer page for \code{monyouth}}.
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{givart}},
\code{\link{giveduc}},
\code{\link{givenvir}},
\code{\link{givfound}},
\code{\link{givhlth}},
\code{\link{givhuman}},
\code{\link{givinfrm}},
\code{\link{givintl}},
\code{\link{givoth}},
\code{\link{givpol}},
\code{\link{givpub}},
\code{\link{givrec}},
\code{\link{givrelig}},
\code{\link{givwork}},
\code{\link{givyouth}},
\code{\link{homeless}},
\code{\link{hrsart}},
\code{\link{hrseduc}},
\code{\link{hrsenvir}},
\code{\link{hrsfound}},
\code{\link{hrshlth}},
\code{\link{hrshuman}},
\code{\link{hrsinfrm}},
\code{\link{hrsintl}},
\code{\link{hrsoth}},
\code{\link{hrspol}},
\code{\link{hrspub}},
\code{\link{hrsrec}},
\code{\link{hrsrelig}},
\code{\link{hrswork}},
\code{\link{hrsyouth}},
\code{\link{monart}},
\code{\link{moneduc}},
\code{\link{monenvir}},
\code{\link{monfound}},
\code{\link{monhlth}},
\code{\link{monhuman}},
\code{\link{moninfrm}},
\code{\link{monintl}},
\code{\link{monoth}},
\code{\link{monpol}},
\code{\link{monpub}},
\code{\link{monrec}},
\code{\link{monrelig}},
\code{\link{monwork}},
\code{\link{needyfrd}},
\code{\link{needynei}},
\code{\link{needyoth}},
\code{\link{needyrel}},
\code{\link{totart}},
\code{\link{toteduc}},
\code{\link{totenvir}},
\code{\link{totfound}},
\code{\link{tothlth}},
\code{\link{tothuman}},
\code{\link{totinfrm}},
\code{\link{totintl}},
\code{\link{tototh}},
\code{\link{totpol}},
\code{\link{totpub}},
\code{\link{totrec}},
\code{\link{totrelig}},
\code{\link{totwork}},
\code{\link{totyouth}},
\code{\link{valart}},
\code{\link{valeduc}},
\code{\link{valenvir}},
\code{\link{valfound}},
\code{\link{valhlth}},
\code{\link{valhuman}},
\code{\link{valinfrm}},
\code{\link{valintl}},
\code{\link{valoth}},
\code{\link{valpol}},
\code{\link{valpub}},
\code{\link{valrec}},
\code{\link{valrelig}},
\code{\link{valwork}},
\code{\link{valyouth}},
\code{\link{volart}},
\code{\link{voleduc}},
\code{\link{volenvir}},
\code{\link{volfound}},
\code{\link{volhlth}},
\code{\link{volhuman}},
\code{\link{volinfrm}},
\code{\link{volintl}},
\code{\link{voloth}},
\code{\link{volpol}},
\code{\link{volpub}},
\code{\link{volrec}},
\code{\link{volrelig}},
\code{\link{volwork}},
\code{\link{volyouth}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
