% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_268.R
\name{moral12}
\alias{moral12}
\title{Past 12 months r web to learn moral issues}
\source{
General Social Survey https://gss.norc.org
}
\description{
moral12
}
\details{
Question In the past 12 months, have you used the Web to find out about or discuss?
C. Morality/family issues?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2430/vshow}{GSS Data Explorer page for \code{moral12}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2453 \tab 72 \tab 208 \tab 84 \tab 2817 \cr
   2002 \tab 2411 \tab 92 \tab 217 \tab 45 \tab 2765 \cr
   Total \tab 4864 \tab 164 \tab 425 \tab 129 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{enviro12}},
\code{\link{fe12}},
\code{\link{foraff12}},
\code{\link{gun12}},
\code{\link{news12}},
\code{\link{polagree}},
\code{\link{polcam12}},
\code{\link{racrel12}},
\code{\link{taxes12}}
}
\concept{Information Society}
\keyword{variable}
