% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_340.R
\name{moregood}
\alias{moregood}
\title{I expect more good things to happen to me than bad}
\source{
General Social Survey https://gss.norc.org
}
\description{
moregood
}
\details{
Question Indicate your agreement with each of the following statements by selecting the number that comes closest to your answer:
I. Overall, I expect more good things to happen to me than bad.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3044/vshow}{GSS Data Explorer page for \code{moregood}}.
}

\seealso{
Other Genes and the Environment: 
\code{\link{athletic}},
\code{\link{depndabl}},
\code{\link{geneexps}},
\code{\link{intrwght}},
\code{\link{kindpers}},
\code{\link{owndoing}},
\code{\link{sadblue}},
\code{\link{selfish}}

Other Satisfaction: 
\code{\link{athletic}},
\code{\link{depndabl}},
\code{\link{kindpers}},
\code{\link{owndoing}},
\code{\link{sadblue}},
\code{\link{selfish}}
}
\concept{Feelings}
\concept{Genes and the Environment}
\concept{Satisfaction}
\concept{Work Organizations}
\keyword{variable}
