% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_424.R
\name{moretrde}
\alias{moretrde}
\title{More trade, less jobs in usa}
\source{
General Social Survey https://gss.norc.org
}
\description{
moretrde
}
\details{
Question Suppose two friends of yours are talking about trade with other countries. One friend thinks that increased trade has created a lot of jobs in the United States for people with jobs like yours. The other friend thinks that increased trade has taken away a lot of jobs in the United States for people with jobs like yours. Which view is closer to your own?
}
\section{Values}{

\itemize{
\item \code{1} created more jobs
\item \code{2} about the same
\item \code{3} taken them away
\item \code{4} not relevant for my job
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3884/vshow}{GSS Data Explorer page for \code{moretrde}}.
}

\seealso{
Other Global Economics: 
\code{\link{findnwjb}},
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{newjobs}},
\code{\link{whyjbct2}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{wkcomptr}}
}
\concept{Global Economics}
\keyword{variable}
