% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_424.R
\name{moretrde}
\alias{moretrde}
\title{More trade, less jobs in usa}
\source{
General Social Survey https://gss.norc.org
}
\description{
moretrde
}
\details{
Question Suppose two friends of yours are talking about trade with other countries. One friend thinks that increased trade has created a lot of jobs in the United States for people with jobs like yours. The other friend thinks that increased trade has taken away a lot of jobs in the United States for people with jobs like yours. Which view is closer to your own?
}
\section{Values}{

\itemize{
\item \verb{[1]} created more jobs
\item \verb{[2]} about the same
\item \verb{[3]} taken them away
\item \verb{[4]} not relevant for my job
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3884/vshow}{GSS Data Explorer page for \code{moretrde}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab created more jobs \tab about the same \tab taken them away \tab not relevant for my job \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1165 \tab 136 \tab 159 \tab 276 \tab 254 \tab 19 \tab 14 \tab 2023 \cr
   Total \tab 1165 \tab 136 \tab 159 \tab 276 \tab 254 \tab 19 \tab 14 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{findnwjb}},
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{newjobs}},
\code{\link{whyjbct2}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{wkcomptr}}
}
\concept{Global Economics}
\keyword{variable}
