% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_681.R
\name{morextrm}
\alias{morextrm}
\title{Everyone is like x but x's case is just a bit more extreme and ongoing}
\source{
General Social Survey https://gss.norc.org
}
\description{
morextrm
}
\details{
Question Let’s continue talking about (Name) for a minute. Tell me how much you agree or disagree with the following statement. Basically, we are all a little like (Name). In (Name)’s case it is just a bit more extreme and ongoing.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8044/vshow}{GSS Data Explorer page for \code{morextrm}}.
}

\seealso{
Other Mental Health: 
\code{\link{bullied}},
\code{\link{bullpeer}},
\code{\link{imprvown1}},
\code{\link{mustdoc1}},
\code{\link{musthosp1}},
\code{\link{mustmed1}},
\code{\link{physill1}},
\code{\link{smedia}}
}
\concept{Children}
\concept{Mental Health}
\keyword{variable}
