% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_496.R
\name{mortravl}
\alias{mortravl}
\title{I would accept a longer commute}
\source{
General Social Survey https://gss.norc.org
}
\description{
mortravl
}
\details{
Question (HAND CARD D22)            
(To what extent do you agree or disagree with the following statements? In order to avoid unemployment I would be willing...)            
To travel longer to get to work.            
(Would you say...)
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6346/vshow}{GSS Data Explorer page for \code{mortravl}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2533 \tab 16 \tab 99 \tab 43 \tab 109 \tab 42 \tab 2 \tab 23 \tab 2867 \cr
   Total \tab 2533 \tab 16 \tab 99 \tab 43 \tab 109 \tab 42 \tab 2 \tab 23 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{exjobsat}},
\code{\link{extraern}},
\code{\link{extrawrk}},
\code{\link{jbendmo}},
\code{\link{jbendyr}},
\code{\link{lowpay}},
\code{\link{newskill}},
\code{\link{tempwork}},
\code{\link{wornojob}}
}
\concept{Work Orientation}
\keyword{variable}
