% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_110.R
\name{mostact}
\alias{mostact}
\title{Which group is r most active in}
\source{
General Social Survey https://gss.norc.org
}
\description{
mostact
}
\details{
Question Which one of these groups or organizations (READ GROUP FROM MEMFRAT THAT R REPLIED "YES" TO) are you most active in?  CODE ONLY ONE.
}
\section{Values}{

\itemize{
\item \code{1} fraternal groups
\item \code{2} service clubs
\item \code{3} veterans groups
\item \code{4} political clubs
\item \code{5} labor unions
\item \code{6} sports groups
\item \code{7} youth groups
\item \code{8} school service groups
\item \code{9} hobby or garden clubs
\item \code{10} school fraternities or sororities
\item \code{11} nationality groups
\item \code{12} farm organizations
\item \code{13} literary art discussion or study grps
\item \code{14} professional or academic societies
\item \code{15} church-affiliated groups
\item \code{16} other groups
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1013/vshow}{GSS Data Explorer page for \code{mostact}}.
}

\seealso{
Other Socio-political participation: 
\code{\link{actchurh}},
\code{\link{actfarm}},
\code{\link{actfrat}},
\code{\link{actgreek}},
\code{\link{acthobby}},
\code{\link{actlit}},
\code{\link{actnat}},
\code{\link{actother}},
\code{\link{actpolit}},
\code{\link{actprof}},
\code{\link{actschl}},
\code{\link{actserv}},
\code{\link{actsport}},
\code{\link{actunion}},
\code{\link{actvet}},
\code{\link{actyouth}},
\code{\link{attndgrp}},
\code{\link{blkgains}},
\code{\link{blkinflu}},
\code{\link{caregrp}},
\code{\link{churhgrp}},
\code{\link{fedtrust}},
\code{\link{fegains}},
\code{\link{feinflu}},
\code{\link{gavepol}},
\code{\link{givegrp}},
\code{\link{governor}},
\code{\link{intpol}},
\code{\link{leadgrp}},
\code{\link{lobbygrp}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{locinflu}},
\code{\link{loclived}},
\code{\link{loclobby}},
\code{\link{locprob}},
\code{\link{locself}},
\code{\link{loctrust}},
\code{\link{locvote}},
\code{\link{memchurh}},
\code{\link{memfarm}},
\code{\link{memfrat}},
\code{\link{memgreek}},
\code{\link{memhobby}},
\code{\link{memlit}},
\code{\link{memnat}},
\code{\link{memnum}},
\code{\link{memother}},
\code{\link{mempolit}},
\code{\link{memprof}},
\code{\link{memschl}},
\code{\link{memserv}},
\code{\link{memsport}},
\code{\link{memunion}},
\code{\link{memvet}},
\code{\link{memyouth}},
\code{\link{othlobby}},
\code{\link{othself}},
\code{\link{polrally}},
\code{\link{schlhead}},
\code{\link{servegrp}},
\code{\link{solchurh}},
\code{\link{solfarm}},
\code{\link{solfrat}},
\code{\link{solgreek}},
\code{\link{solhobby}},
\code{\link{sollit}},
\code{\link{solnat}},
\code{\link{solother}},
\code{\link{solpolit}},
\code{\link{solprof}},
\code{\link{solschl}},
\code{\link{solserv}},
\code{\link{solsport}},
\code{\link{solunion}},
\code{\link{solvet}},
\code{\link{solyouth}},
\code{\link{swayvote}},
\code{\link{usrep}},
\code{\link{workpol}},
\code{\link{writegrp}}
}
\concept{Socio-political participation}
\keyword{variable}
