% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_260.R
\name{mostcom}
\alias{mostcom}
\title{Have most in common with which group}
\source{
General Social Survey https://gss.norc.org
}
\description{
mostcom
}
\details{
Question Of these groups–whites, blacks, Jews, Hispanics and Latin Americans, Asian Americans, if you had to say, which One (other than your own) do you feel you have the most in common with?
}
\section{Values}{

\itemize{
\item \code{1} whites
\item \code{2} blacks/african-americans
\item \code{3} jews
\item \code{4} hispanics
\item \code{5} asians
\item \code{6} equal in common to all
\item \code{7} nothing in common with any
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2360/vshow}{GSS Data Explorer page for \code{mostcom}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab whites \tab blacks/african-americans \tab jews \tab hispanics \tab asians \tab equal in common to all \tab nothing in common with any \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1419 \tab 286 \tab 213 \tab 180 \tab 197 \tab 99 \tab 221 \tab 97 \tab 75 \tab 30 \tab 2817 \cr
   Total \tab 1419 \tab 286 \tab 213 \tab 180 \tab 197 \tab 99 \tab 221 \tab 97 \tab 75 \tab 30 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{asncls}},
\code{\link{asncom}},
\code{\link{asnrel}},
\code{\link{asnschl}},
\code{\link{asnwrk}},
\code{\link{hispcls}},
\code{\link{hispcom}},
\code{\link{hisprel}},
\code{\link{hispwrk}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
