% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_303.R
\name{mosthrs}
\alias{mosthrs}
\title{Most hrs/week worked in past month}
\source{
General Social Survey https://gss.norc.org
}
\description{
mosthrs
}
\details{
Question In the last month, what is the greatest number of hours you worked in a week, at all paid jobs? Please consider all hours, including any extra hours, overtime, work you did at home for your job, and time you spent on work that may not have been directly billable or compensated.
HOURS:
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6293/vshow}{GSS Data Explorer page for \code{mosthrs}}.
}

\seealso{
Other Flexible Work: 
\code{\link{paidhow}},
\code{\link{usualhrs}}
}
\concept{Flexible Work}
\keyword{variable}
