% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_154.R
\name{mostimp4}
\alias{mostimp4}
\title{4th imp event in lst 60 yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
mostimp4
}
\details{
Question The next questions concern how people think about the past. There have been a lot of national and world events  and changes over the past 60 years--say, from about 1930 right up until today. Would you mention one or two such events or changes that seem to you to have been especially important? There aren't any right or wrong answers to the question--just whatever national or world events or changes over the past 60 years that come to mind as important to you.
}
\section{Values}{

\itemize{
\item \code{990} other
\item \code{995} personal
\item \code{996} none, no event
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1449/vshow}{GSS Data Explorer page for \code{mostimp4}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 18 \tab 19 \tab 2 \tab 21 \tab 26 \tab 29 \tab 3 \tab 31 \tab 32 \tab 34 \tab 39 \tab 40 \tab 42 \tab 43 \tab 44 \tab 50 \tab 51 \tab 52 \tab 55 \tab 6 \tab 64 \tab 71 \tab 73 \tab 75 \tab 77 \tab 79 \tab other \tab personal \tab don't know \tab no answer \tab Total \cr
   1993 \tab 1447 \tab 1 \tab 6 \tab 2 \tab 3 \tab 2 \tab 3 \tab 1 \tab 3 \tab 3 \tab 7 \tab 2 \tab 7 \tab 16 \tab 6 \tab 2 \tab 3 \tab 12 \tab 3 \tab 1 \tab 22 \tab 5 \tab 1 \tab 5 \tab 9 \tab 1 \tab 1 \tab 1 \tab 8 \tab 1 \tab 1 \tab 21 \tab 1606 \cr
   Total \tab 1447 \tab 1 \tab 6 \tab 2 \tab 3 \tab 2 \tab 3 \tab 1 \tab 3 \tab 3 \tab 7 \tab 2 \tab 7 \tab 16 \tab 6 \tab 2 \tab 3 \tab 12 \tab 3 \tab 1 \tab 22 \tab 5 \tab 1 \tab 5 \tab 9 \tab 1 \tab 1 \tab 1 \tab 8 \tab 1 \tab 1 \tab 21 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{colmajr1}},
\code{\link{colmajr2}},
\code{\link{hsclass1}},
\code{\link{hsclass2}}
}
\concept{Culture}
\keyword{variable}
