% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_154.R
\name{mostimp4}
\alias{mostimp4}
\title{4th imp event in lst 60 yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
mostimp4
}
\details{
Question The next questions concern how people think about the past. There have been a lot of national and world events  and changes over the past 60 years--say, from about 1930 right up until today. Would you mention one or two such events or changes that seem to you to have been especially important? There aren't any right or wrong answers to the question--just whatever national or world events or changes over the past 60 years that come to mind as important to you.
}
\section{Values}{

\itemize{
\item \code{990} other
\item \code{995} personal
\item \code{996} none, no event
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1449/vshow}{GSS Data Explorer page for \code{mostimp4}}.
}

\seealso{
Other Culture: 
\code{\link{colmajr1}},
\code{\link{colmajr2}},
\code{\link{hsclass1}},
\code{\link{hsclass2}}
}
\concept{Culture}
\keyword{variable}
