% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_283.R
\name{mostloc}
\alias{mostloc}
\title{Where does r use computer most}
\source{
General Social Survey https://gss.norc.org
}
\description{
mostloc
}
\details{
Question Which location do you use the most?
}
\section{Values}{

\itemize{
\item \verb{[1]} school
\item \verb{[2]} library
\item \verb{[3]} friends house
\item \verb{[4]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2584/vshow}{GSS Data Explorer page for \code{mostloc}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab school \tab library \tab friends house \tab other \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2727 \tab 23 \tab 10 \tab 19 \tab 7 \tab 31 \tab - \tab 2817 \cr
   2002 \tab 2662 \tab 30 \tab 26 \tab 31 \tab 12 \tab 3 \tab 1 \tab 2765 \cr
   Total \tab 5389 \tab 53 \tab 36 \tab 50 \tab 19 \tab 34 \tab 1 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{compelse}},
\code{\link{compfri}},
\code{\link{complib}},
\code{\link{compoth}},
\code{\link{compschl}},
\code{\link{ecomhro}},
\code{\link{ecommino}},
\code{\link{perhro}},
\code{\link{permino}}
}
\concept{Information Society}
\keyword{variable}
