% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_497.R
\name{moveaway}
\alias{moveaway}
\title{I would move outside the usa for work}
\source{
General Social Survey https://gss.norc.org
}
\description{
moveaway
}
\details{
Question (HAND CARD D22)            
(To what extent do you agree or disagree with the following statements? In order to avoid unemployment I would be willing...)            
To move to a different country.            
(Would you say...)
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6348/vshow}{GSS Data Explorer page for \code{moveaway}}.
}

\seealso{
Other Work Orientation: 
\code{\link{econsup1}},
\code{\link{moveinus}},
\code{\link{unfamten}},
\code{\link{unjobexp}},
\code{\link{unmoney}},
\code{\link{unpeople}},
\code{\link{unrespct}},
\code{\link{unslfcon}},
\code{\link{worecsup}}
}
\concept{Work Orientation}
\keyword{variable}
