% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_497.R
\name{moveinus}
\alias{moveinus}
\title{I would move within the usa for work}
\source{
General Social Survey https://gss.norc.org
}
\description{
moveinus
}
\details{
Question (HAND CARD D22)            
(To what extent do you agree or disagree with the following statements? In order to avoid unemployment I would be willing...)            
To move within America.            
(Would you say...)
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6347/vshow}{GSS Data Explorer page for \code{moveinus}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2533 \tab 22 \tab 90 \tab 35 \tab 86 \tab 75 \tab 3 \tab 23 \tab 2867 \cr
   Total \tab 2533 \tab 22 \tab 90 \tab 35 \tab 86 \tab 75 \tab 3 \tab 23 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{econsup1}},
\code{\link{moveaway}},
\code{\link{unfamten}},
\code{\link{unjobexp}},
\code{\link{unmoney}},
\code{\link{unpeople}},
\code{\link{unrespct}},
\code{\link{unslfcon}},
\code{\link{worecsup}}
}
\concept{Work Orientation}
\keyword{variable}
