% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_529.R
\name{movestat}
\alias{movestat}
\title{Willing to move to another state}
\source{
General Social Survey https://gss.norc.org
}
\description{
movestat
}
\details{
Question If you could improve your work or living conditions, how willing or unwilling would you be to…
C. Move to another state
}
\section{Values}{

\itemize{
\item \verb{[1]} very willing
\item \verb{[2]} fairly willing
\item \verb{[3]} neither willing nor unwilling
\item \verb{[4]} fairly unwilling
\item \verb{[5]} very unwilling
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4823/vshow}{GSS Data Explorer page for \code{movestat}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very willing \tab fairly willing \tab neither willing nor unwilling \tab fairly unwilling \tab very unwilling \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 271 \tab 303 \tab 190 \tab 218 \tab 320 \tab 39 \tab 26 \tab 2904 \cr
   Total \tab 1537 \tab 271 \tab 303 \tab 190 \tab 218 \tab 320 \tab 39 \tab 26 \tab 2904 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{ambornin}},
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{clseusa}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movetown}},
\code{\link{moveusa}},
\code{\link{onenatn}}
}
\concept{ISSP National Identity}
\keyword{variable}
