% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_529.R
\name{moveusa}
\alias{moveusa}
\title{Willing to move outside america}
\source{
General Social Survey https://gss.norc.org
}
\description{
moveusa
}
\details{
Question If you could improve your work or living conditions, how willing or unwilling would you be to…
D. Move outside America
}
\section{Values}{

\itemize{
\item \verb{[1]} very willing
\item \verb{[2]} fairly willing
\item \verb{[3]} neither willing nor unwilling
\item \verb{[4]} fairly unwilling
\item \verb{[5]} very unwilling
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4824/vshow}{GSS Data Explorer page for \code{moveusa}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very willing \tab fairly willing \tab neither willing nor unwilling \tab fairly unwilling \tab very unwilling \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 83 \tab 123 \tab 149 \tab 208 \tab 733 \tab 45 \tab 26 \tab 2904 \cr
   Total \tab 1537 \tab 83 \tab 123 \tab 149 \tab 208 \tab 733 \tab 45 \tab 26 \tab 2904 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{ambornin}},
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{clseusa}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{movetown}},
\code{\link{onenatn}}
}
\concept{ISSP National Identity}
\keyword{variable}
