% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_291.R
\name{mp3}
\alias{mp3}
\title{How familiar r is with term mp3}
\source{
General Social Survey https://gss.norc.org
}
\description{
mp3
}
\details{
Question F. MP3
}
\section{Values}{

\itemize{
\item \code{1} very familiar
\item \code{2} somewhat familiar
\item \code{3} not familiar
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2635/vshow}{GSS Data Explorer page for \code{mp3}}.
}

\seealso{
Other Information Society: 
\code{\link{advsrch}},
\code{\link{browser3}},
\code{\link{ezines}},
\code{\link{newsgrps}},
\code{\link{prefsets}},
\code{\link{timekid1}},
\code{\link{timekid2}},
\code{\link{timekid3}},
\code{\link{timekid4}}

Other Computers And Internet: 
\code{\link{advsrch}},
\code{\link{browser3}},
\code{\link{ezines}},
\code{\link{newsgrps}},
\code{\link{prefsets}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
