% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_483.R
\name{mrmom}
\alias{mrmom}
\title{Gender role reversal hurts family}
\source{
General Social Survey https://gss.norc.org
}
\description{
mrmom
}
\details{
Question And, do you agree or disagree . . .
B. It is not good if the man stays at home and cares for the children and the woman goes out to work.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4403/vshow}{GSS Data Explorer page for \code{mrmom}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1994 \tab 1545 \tab 99 \tab 231 \tab 291 \tab 560 \tab 224 \tab 16 \tab 26 \tab 2992 \cr
   Total \tab 1545 \tab 99 \tab 231 \tab 291 \tab 560 \tab 224 \tab 16 \tab 26 \tab 2992 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{cohabfst}},
\code{\link{cohabok}},
\code{\link{divbest}},
\code{\link{divifkd1}},
\code{\link{divnokd1}},
\code{\link{fewrksup}},
\code{\link{hubbywk1}},
\code{\link{meovrwrk}},
\code{\link{singlpar}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
