% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_357.R
\name{mtfrat1}
\alias{mtfrat1}
\title{How many years r born a member of fraternity group1}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtfrat1
}
\details{
Question A. Does the fraternal club meet in this area?
\enumerate{
\item First fraternal club
}
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3221/vshow}{GSS Data Explorer page for \code{mtfrat1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2707 \tab 77 \tab 19 \tab 1 \tab 8 \tab 2812 \cr
   Total \tab 2707 \tab 77 \tab 19 \tab 1 \tab 8 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtfrat2}},
\code{\link{mtfrat3}},
\code{\link{yrinfrm1}},
\code{\link{yrinfrm2}},
\code{\link{yrinfrm3}},
\code{\link{yrinfrm4}},
\code{\link{yrinfrm5}},
\code{\link{yrother4}},
\code{\link{yrother5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
