% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_357.R
\name{mtfrat2}
\alias{mtfrat2}
\title{How many years r born a member of fraternity group2}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtfrat2
}
\details{
Question A. Does the fraternal club meet in this area?
2. Second fraternal club
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3222/vshow}{GSS Data Explorer page for \code{mtfrat2}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2004 \tab 2787 \tab 15 \tab 3 \tab 7 \tab 2812 \cr
   Total \tab 2787 \tab 15 \tab 3 \tab 7 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtfrat1}},
\code{\link{mtfrat3}},
\code{\link{yrinfrm1}},
\code{\link{yrinfrm2}},
\code{\link{yrinfrm3}},
\code{\link{yrinfrm4}},
\code{\link{yrinfrm5}},
\code{\link{yrother4}},
\code{\link{yrother5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
