% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_362.R
\name{mtnat3}
\alias{mtnat3}
\title{Nationality group3 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtnat3
}
\details{
Question K. Does the nationality group meet in this area?
3. Third nationality group
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3268/vshow}{GSS Data Explorer page for \code{mtnat3}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtfarm1}},
\code{\link{mtfarm2}},
\code{\link{mtfarm3}},
\code{\link{mtgreek2}},
\code{\link{mtgreek3}},
\code{\link{mtlit1}},
\code{\link{mtlit2}},
\code{\link{mtnat1}},
\code{\link{mtnat2}}

Other Voluntary Associations: 
\code{\link{mtfarm1}},
\code{\link{mtfarm2}},
\code{\link{mtfarm3}},
\code{\link{mtgreek2}},
\code{\link{mtgreek3}},
\code{\link{mtlit1}},
\code{\link{mtlit2}},
\code{\link{mtnat1}},
\code{\link{mtnat2}}
}
\concept{Voluntary Assocations/Social Networks}
\concept{Voluntary Associations}
\keyword{variable}
