% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_360.R
\name{mtsport3}
\alias{mtsport3}
\title{Sport club3 meets in this area}
\source{
General Social Survey https://gss.norc.org
}
\description{
mtsport3
}
\details{
Question F. Does the sports group meet in this area?
3. Third sports group
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3247/vshow}{GSS Data Explorer page for \code{mtsport3}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtsport1}},
\code{\link{mtsport2}},
\code{\link{mtsport4}},
\code{\link{mtsport5}},
\code{\link{mtsport6}},
\code{\link{mtunion4}},
\code{\link{mtyouth1}},
\code{\link{mtyouth2}},
\code{\link{mtyouth3}}

Other Voluntary Associations: 
\code{\link{mtsport1}},
\code{\link{mtsport2}},
\code{\link{mtsport4}},
\code{\link{mtsport5}},
\code{\link{mtsport6}},
\code{\link{mtunion4}},
\code{\link{mtyouth1}},
\code{\link{mtyouth2}},
\code{\link{mtyouth3}}
}
\concept{Voluntary Assocations/Social Networks}
\concept{Voluntary Associations}
\keyword{variable}
