% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_271.R
\name{musicbuy}
\alias{musicbuy}
\title{Use web to buy music after hearing on web}
\source{
General Social Survey https://gss.norc.org
}
\description{
musicbuy
}
\details{
Question E. Purchased a recording as a result of hearing it on the Web?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2460/vshow}{GSS Data Explorer page for \code{musicbuy}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2498 \tab 75 \tab 131 \tab 113 \tab 2817 \cr
   2002 \tab 2492 \tab 100 \tab 127 \tab 46 \tab 2765 \cr
   Total \tab 4990 \tab 175 \tab 258 \tab 159 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{musiclst}},
\code{\link{wwwclass}},
\code{\link{wwwcntry}},
\code{\link{wwwgthic}},
\code{\link{wwwjazz}},
\code{\link{wwwoldie}},
\code{\link{wwwrap}},
\code{\link{wwwrelig}},
\code{\link{wwwrock}}
}
\concept{Information Society}
\keyword{variable}
