% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_272.R
\name{musicnew}
\alias{musicnew}
\title{Visit sites about music not usually heard}
\source{
General Social Survey https://gss.norc.org
}
\description{
musicnew
}
\details{
Question Is it VERY TRUE for you, SOMEWHAT TRUE, or NOT TRUE that:
C. You visit sites where you can learn about or hear types of music that you wouldn’t ordinarily listen to?
}
\section{Values}{

\itemize{
\item \code{1} very true
\item \code{2} somewhat true
\item \code{3} not true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2472/vshow}{GSS Data Explorer page for \code{musicnew}}.
}

\seealso{
Other Information Society: 
\code{\link{artinfo}},
\code{\link{artmus1}},
\code{\link{artmus2}},
\code{\link{artmus3}},
\code{\link{artview}},
\code{\link{musicdif}},
\code{\link{musiclke}},
\code{\link{wwwart}},
\code{\link{wwwworld}}
}
\concept{Information Society}
\keyword{variable}
