% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_272.R
\name{musicnew}
\alias{musicnew}
\title{Visit sites about music not usually heard}
\source{
General Social Survey https://gss.norc.org
}
\description{
musicnew
}
\details{
Question Is it VERY TRUE for you, SOMEWHAT TRUE, or NOT TRUE that:
C. You visit sites where you can learn about or hear types of music that you wouldn’t ordinarily listen to?
}
\section{Values}{

\itemize{
\item \code{1} very true
\item \code{2} somewhat true
\item \code{3} not true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2472/vshow}{GSS Data Explorer page for \code{musicnew}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not true \tab no answer \tab Total \cr
   2000 \tab 2498 \tab 53 \tab 60 \tab 91 \tab 115 \tab 2817 \cr
   2002 \tab 2492 \tab 48 \tab 80 \tab 99 \tab 46 \tab 2765 \cr
   Total \tab 4990 \tab 101 \tab 140 \tab 190 \tab 161 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{artinfo}},
\code{\link{artmus1}},
\code{\link{artmus2}},
\code{\link{artmus3}},
\code{\link{artview}},
\code{\link{musicdif}},
\code{\link{musiclke}},
\code{\link{wwwart}},
\code{\link{wwwworld}}
}
\concept{Information Society}
\keyword{variable}
