% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_440.R
\name{mustret}
\alias{mustret}
\title{Govt should enforce mandatory retirement}
\source{
General Social Survey https://gss.norc.org
}
\description{
mustret
}
\details{
Question Do you think that...
C. All employees should be required to retire at an age set by law.
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4038/vshow}{GSS Data Explorer page for \code{mustret}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{poleff1}},
\code{\link{poleff2}},
\code{\link{poleff3}},
\code{\link{poleff4}},
\code{\link{poleff5}},
\code{\link{poleff6}},
\code{\link{poleff7}},
\code{\link{poleff8}},
\code{\link{poleff9}}
}
\concept{Federal Government}
\concept{ISSP Role of Gov't}
\keyword{variable}
