% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_466.R
\name{mygoals}
\alias{mygoals}
\title{It is easy for r to accomplish their goals}
\source{
General Social Survey https://gss.norc.org
}
\description{
mygoals
}
\details{
Question To what extent is the following statement true or untrue for you? It is easy for me to accomplish my goals
}
\section{Values}{

\itemize{
\item \code{1} completely true
\item \code{2} mostly true
\item \code{3} somewhat true
\item \code{4} neither true nor untrue
\item \code{5} somewhat untrue
\item \code{6} mostly untrue
\item \code{7} completely untrue
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7274/vshow}{GSS Data Explorer page for \code{mygoals}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab completely true \tab mostly true \tab somewhat true \tab neither true nor untrue \tab somewhat untrue \tab mostly untrue \tab completely untrue \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1175 \tab 143 \tab 475 \tab 345 \tab 104 \tab 60 \tab 29 \tab 11 \tab 5 \tab 1 \tab 2348 \cr
   Total \tab 1175 \tab 143 \tab 475 \tab 345 \tab 104 \tab 60 \tab 29 \tab 11 \tab 5 \tab 1 \tab 2348 \cr
}
}

\keyword{variable}
