% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_429.R
\name{mykids}
\alias{mykids}
\title{R's current insurance plan covers children}
\source{
General Social Survey https://gss.norc.org
}
\description{
mykids
}
\details{
Question Are the following people in your family covered by your current insurance plan?
B. Children for whom I am a biological or adoptive parent
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} not applicable
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3933/vshow}{GSS Data Explorer page for \code{mykids}}.
}

\seealso{
Other Sexual Orientation: 
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{hlthcovr}},
\code{\link{hlthtype}},
\code{\link{othkids}},
\code{\link{sppart}}

Other Children: 
\code{\link{othkids}}

Other Family: 
\code{\link{othkids}},
\code{\link{sppart}}

Other Health: 
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{hlthcovr}},
\code{\link{hlthtype}},
\code{\link{othkids}},
\code{\link{sppart}}

Other Job: 
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{harsexjb}},
\code{\link{hlthtype}}
}
\concept{Children}
\concept{Family}
\concept{Health}
\concept{Job}
\concept{Sexual Orientation}
\keyword{variable}
