% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_224.R
\name{myprobs2}
\alias{myprobs2}
\title{How much person 2's mental health condition causes r problems}
\source{
General Social Survey https://gss.norc.org
}
\description{
myprobs2
}
\details{
Question On a scale from 1 to 10, with 1 being not at all and 10 being very much, how much does (NAME 2)'s mental health condition cause you problems, create stress, or make your life difficult?
}
\section{Values}{

\itemize{
\item \code{1} 1 - not at all
\item \code{10} 10 - very much
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7126/vshow}{GSS Data Explorer page for \code{myprobs2}}.
}

\keyword{variable}
