% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_516.R
\name{mywaygod}
\alias{mywaygod}
\title{R has own way of connecting with god without churches or religious services}
\source{
General Social Survey https://gss.norc.org
}
\description{
mywaygod
}
\details{
Question Do you agree or disagree with the following...I have my own way of connecting with God without churches or religious services.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2018 \cr
   B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4708/vshow}{GSS Data Explorer page for \code{mywaygod}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2008 \tab 658 \tab 339 \tab 543 \tab 155 \tab 194 \tab 118 \tab 9 \tab 7 \tab 2023 \cr
   2018 \tab 1173 \tab 299 \tab 443 \tab 144 \tab 158 \tab 114 \tab 8 \tab 9 \tab 2348 \cr
   Total \tab 1831 \tab 638 \tab 986 \tab 299 \tab 352 \tab 232 \tab 17 \tab 16 \tab 4371 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{ancestrs}},
\code{\link{nirvana}},
\code{\link{reincar}},
\code{\link{relcand}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relmarry}}
}
\concept{ISSP Religion}
\keyword{variable}
