% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_516.R
\name{mywaygod}
\alias{mywaygod}
\title{R has own way of connecting with god without churches or religious services}
\source{
General Social Survey https://gss.norc.org
}
\description{
mywaygod
}
\details{
Question Do you agree or disagree with the following...I have my own way of connecting with God without churches or religious services.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab B/C \tab partial \cr
   2018 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4708/vshow}{GSS Data Explorer page for \code{mywaygod}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{ancestrs}},
\code{\link{nirvana}},
\code{\link{reincar}},
\code{\link{relcand}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relmarry}}

Other Religion: 
\code{\link{ancestrs}},
\code{\link{nirvana}},
\code{\link{reincar}},
\code{\link{relcand}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relmarry}}
}
\concept{ISSP Religion}
\concept{Religion}
\keyword{variable}
