% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_534.R
\name{nafta1}
\alias{nafta1}
\title{How much heard or read about nafta?}
\source{
General Social Survey https://gss.norc.org
}
\description{
nafta1
}
\details{
Question How much have you heard or read about the North American Free Trade Agreement or NAFTA?
}
\section{Values}{

\itemize{
\item \verb{[1]} a lot
\item \verb{[2]} quite a bit
\item \verb{[3]} not much
\item \verb{[4]} nothing at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4876/vshow}{GSS Data Explorer page for \code{nafta1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a lot \tab quite a bit \tab not much \tab nothing at all \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 181 \tab 406 \tab 561 \tab 199 \tab 1 \tab 19 \tab 2904 \cr
   2004 \tab 1596 \tab 143 \tab 289 \tab 554 \tab 221 \tab 3 \tab 6 \tab 2812 \cr
   2014 \tab 1264 \tab 94 \tab 223 \tab 576 \tab 337 \tab 40 \tab 4 \tab 2538 \cr
   Total \tab 4397 \tab 418 \tab 918 \tab 1691 \tab 757 \tab 44 \tab 29 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{citizen}},
\code{\link{ethclose}},
\code{\link{nafta2}},
\code{\link{parcit}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang1}},
\code{\link{spklang2}},
\code{\link{spklang3}}
}
\concept{ISSP National Identity}
\keyword{variable}
