% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_534.R
\name{nafta1}
\alias{nafta1}
\title{How much heard or read about nafta?}
\source{
General Social Survey https://gss.norc.org
}
\description{
nafta1
}
\details{
Question How much have you heard or read about the North American Free Trade Agreement or NAFTA?
}
\section{Values}{

\itemize{
\item \code{1} a lot
\item \code{2} quite a bit
\item \code{3} not much
\item \code{4} nothing at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4876/vshow}{GSS Data Explorer page for \code{nafta1}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{citizen}},
\code{\link{ethclose}},
\code{\link{nafta2}},
\code{\link{parcit}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang1}},
\code{\link{spklang2}},
\code{\link{spklang3}}

Other United States: 
\code{\link{citizen}},
\code{\link{nafta2}},
\code{\link{parcit}}
}
\concept{ISSP National Identity}
\concept{United States}
\keyword{variable}
