% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_536.R
\name{nafta2a}
\alias{nafta2a}
\title{America benefits from being a member of nafta?}
\source{
General Social Survey https://gss.norc.org
}
\description{
nafta2a
}
\details{
Question A. How strongly do you agree or disagree with the following statement? (Please, select one response)
Generally speaking, would you say that America benefits or does not benefit from being a member of NAFTA?
}
\section{Values}{

\itemize{
\item \verb{[1]} greatly benefits
\item \verb{[2]} largely benefits
\item \verb{[3]} somewhat benefits
\item \verb{[4]} benefits only a little
\item \verb{[5]} does not benefit at all
\item \verb{[7]} have never heard of nafta
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4896/vshow}{GSS Data Explorer page for \code{nafta2a}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab greatly benefits \tab largely benefits \tab somewhat benefits \tab benefits only a little \tab does not benefit at all \tab have never heard of nafta \tab don't know \tab no answer \tab Total \cr
   2004 \tab 1596 \tab 52 \tab 105 \tab 360 \tab 144 \tab 104 \tab 109 \tab 332 \tab 10 \tab 2812 \cr
   2014 \tab 1601 \tab 57 \tab 116 \tab 321 \tab 110 \tab 79 \tab 44 \tab 202 \tab 8 \tab 2538 \cr
   Total \tab 3197 \tab 109 \tab 221 \tab 681 \tab 254 \tab 183 \tab 153 \tab 534 \tab 18 \tab 5350 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amproud1}},
\code{\link{forcult}},
\code{\link{immcosts}},
\code{\link{immimp}},
\code{\link{immrghts}},
\code{\link{internet}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{nafta3}}
}
\concept{ISSP National Identity}
\keyword{variable}
