% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_536.R
\name{nafta2a}
\alias{nafta2a}
\title{America benefits from being a member of nafta?}
\source{
General Social Survey https://gss.norc.org
}
\description{
nafta2a
}
\details{
Question A. How strongly do you agree or disagree with the following statement? (Please, select one response)
Generally speaking, would you say that America benefits or does not benefit from being a member of NAFTA?
}
\section{Values}{

\itemize{
\item \code{1} greatly benefits
\item \code{2} largely benefits
\item \code{3} somewhat benefits
\item \code{4} benefits only a little
\item \code{5} does not benefit at all
\item \code{7} have never heard of nafta
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4896/vshow}{GSS Data Explorer page for \code{nafta2a}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{amproud1}},
\code{\link{forcult}},
\code{\link{immcosts}},
\code{\link{immimp}},
\code{\link{immrghts}},
\code{\link{internet}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{nafta3}}

Other United States: 
\code{\link{amproud1}},
\code{\link{forcult}},
\code{\link{nafta3}}
}
\concept{ISSP National Identity}
\concept{United States}
\keyword{variable}
