% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_536.R
\name{nafta3}
\alias{nafta3}
\title{America should follow nafta decision}
\source{
General Social Survey https://gss.norc.org
}
\description{
nafta3
}
\details{
Question B. How strongly do you agree or disagree with the following statement? (Please, select one response)
America should follow NAFTA decisions, even if it does not agree with them.
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} disagree strongly
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4897/vshow}{GSS Data Explorer page for \code{nafta3}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{amproud1}},
\code{\link{forcult}},
\code{\link{immcosts}},
\code{\link{immimp}},
\code{\link{immrghts}},
\code{\link{internet}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{nafta2a}}

Other United States: 
\code{\link{amproud1}},
\code{\link{forcult}},
\code{\link{nafta2a}}
}
\concept{ISSP National Identity}
\concept{United States}
\keyword{variable}
