% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_381.R
\name{nanoben}
\alias{nanoben}
\title{Balance will be strongly in favor of the benefits or slightly}
\source{
General Social Survey https://gss.norc.org
}
\description{
nanoben
}
\details{
Question B. (IF NANOWILL=1) Would you say that the balance will be strongly in favor of the benefits, or only slightly?
}
\section{Values}{

\itemize{
\item \code{1} strongly in favor
\item \code{2} slightly in favor
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/C/- \tab partial \cr
   2016 \tab A/-/- \tab partial \cr
   2018 \tab A/-/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3449/vshow}{GSS Data Explorer page for \code{nanoben}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{intecon}},
\code{\link{inteduc}},
\code{\link{intfarm}},
\code{\link{intintl}},
\code{\link{intmed}},
\code{\link{intsci}},
\code{\link{inttech}},
\code{\link{nanoharm}},
\code{\link{scimode}}

Other Science: 
\code{\link{intmed}},
\code{\link{intsci}},
\code{\link{nanoharm}}
}
\concept{Knowledge}
\concept{Science}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
