% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_380.R
\name{nanoknw2}
\alias{nanoknw2}
\title{True-false: nanoscale differ from same material at larger scale}
\source{
General Social Survey https://gss.norc.org
}
\description{
nanoknw2
}
\details{
Question Here are a couple of true-false questions about nanotechnology. As before, if you don't know or aren't sure, just tell me so, and we will skip to the next question. Remember true, false, or don't know.
B. The properties of nanoscale materials often differ fundamentally and unexpectedly from the properties of the same materials at larger scales. (Is that true or false?)
}
\section{Values}{

\itemize{
\item \verb{[1]} true
\item \verb{[2]} false
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3447/vshow}{GSS Data Explorer page for \code{nanoknw2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab true \tab false \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 3688 \tab 336 \tab 81 \tab 403 \tab 2 \tab 4510 \cr
   2008 \tab 1312 \tab 309 \tab 89 \tab 308 \tab 5 \tab 2023 \cr
   2010 \tab 1796 \tab 77 \tab 32 \tab 130 \tab 9 \tab 2044 \cr
   Total \tab 6796 \tab 722 \tab 202 \tab 841 \tab 16 \tab 8577 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{artseals}},
\code{\link{caremost}},
\code{\link{inuitway}},
\code{\link{nanoknw1}},
\code{\link{nanotech}},
\code{\link{nanowill}},
\code{\link{noicecap}},
\code{\link{penguins}},
\code{\link{sealevel}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
