% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_380.R
\name{nanotech}
\alias{nanotech}
\title{How much r heard about nanotechnology}
\source{
General Social Survey https://gss.norc.org
}
\description{
nanotech
}
\details{
Question How much have you heard about nanotechnology? Have you heard a lot, some, just a little, or nothing at all?
}
\section{Values}{

\itemize{
\item \code{1} a lot
\item \code{2} some
\item \code{3} just a little
\item \code{4} nothing at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C/- \tab partial \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3445/vshow}{GSS Data Explorer page for \code{nanotech}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{artseals}},
\code{\link{caremost}},
\code{\link{inuitway}},
\code{\link{nanoknw1}},
\code{\link{nanoknw2}},
\code{\link{nanowill}},
\code{\link{noicecap}},
\code{\link{penguins}},
\code{\link{sealevel}}

Other Knowledge: 
\code{\link{nanoknw1}},
\code{\link{nanoknw2}},
\code{\link{nanowill}}

Other Science: 
\code{\link{artseals}},
\code{\link{caremost}},
\code{\link{inuitway}},
\code{\link{nanoknw1}},
\code{\link{nanoknw2}},
\code{\link{nanowill}},
\code{\link{noicecap}},
\code{\link{sealevel}}
}
\concept{Knowledge}
\concept{Science}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
