% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_380.R
\name{nanotech}
\alias{nanotech}
\title{How much r heard about nanotechnology}
\source{
General Social Survey https://gss.norc.org
}
\description{
nanotech
}
\details{
Question How much have you heard about nanotechnology? Have you heard a lot, some, just a little, or nothing at all?
}
\section{Values}{

\itemize{
\item \code{1} a lot
\item \code{2} some
\item \code{3} just a little
\item \code{4} nothing at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3445/vshow}{GSS Data Explorer page for \code{nanotech}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a lot \tab some \tab just a little \tab nothing at all \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 100 \tab 284 \tab 436 \tab 1014 \tab 28 \tab 2 \tab 4510 \cr
   2008 \tab 518 \tab 80 \tab 224 \tab 407 \tab 746 \tab 32 \tab 16 \tab 2023 \cr
   2010 \tab 1584 \tab 23 \tab 90 \tab 127 \tab 210 \tab 2 \tab 8 \tab 2044 \cr
   Total \tab 4748 \tab 203 \tab 598 \tab 970 \tab 1970 \tab 62 \tab 26 \tab 8577 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{artseals}},
\code{\link{caremost}},
\code{\link{inuitway}},
\code{\link{nanoknw1}},
\code{\link{nanoknw2}},
\code{\link{nanowill}},
\code{\link{noicecap}},
\code{\link{penguins}},
\code{\link{sealevel}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
