% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_380.R
\name{nanowill}
\alias{nanowill}
\title{Benefit of nanotechnology outweigh harmful results}
\source{
General Social Survey https://gss.norc.org
}
\description{
nanowill
}
\details{
Question A. Nanotechnology works at the molecular level atom by atom to build new structures, materials, and machines. People have frequently noted that new technologies have produced both benefits and harmful results. Do you think the benefits of nanotechnology will outweigh the harmful results or the harmful results will outweigh the benefits?
}
\section{Values}{

\itemize{
\item \verb{[1]} benefits greater
\item \verb{[2]} about equal (if volunteered)
\item \verb{[3]} harmful results greater
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3448/vshow}{GSS Data Explorer page for \code{nanowill}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab benefits greater \tab about equal (if volunteered) \tab harmful results greater \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 727 \tab 355 \tab 166 \tab 614 \tab 2 \tab 4510 \cr
   2008 \tab 518 \tab 577 \tab 171 \tab 148 \tab 582 \tab 27 \tab 2023 \cr
   2010 \tab 1584 \tab 164 \tab 42 \tab 41 \tab 202 \tab 11 \tab 2044 \cr
   2016 \tab 1956 \tab 446 \tab 97 \tab 160 \tab 198 \tab 10 \tab 2867 \cr
   2018 \tab 1563 \tab 421 \tab 63 \tab 163 \tab 128 \tab 10 \tab 2348 \cr
   Total \tab 8267 \tab 2335 \tab 728 \tab 678 \tab 1724 \tab 60 \tab 13792 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{artseals}},
\code{\link{caremost}},
\code{\link{inuitway}},
\code{\link{nanoknw1}},
\code{\link{nanoknw2}},
\code{\link{nanotech}},
\code{\link{noicecap}},
\code{\link{penguins}},
\code{\link{sealevel}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
