% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_380.R
\name{nanowill}
\alias{nanowill}
\title{Benefit of nanotechnology outweigh harmful results}
\source{
General Social Survey https://gss.norc.org
}
\description{
nanowill
}
\details{
Question A. Nanotechnology works at the molecular level atom by atom to build new structures, materials, and machines. People have frequently noted that new technologies have produced both benefits and harmful results. Do you think the benefits of nanotechnology will outweigh the harmful results or the harmful results will outweigh the benefits?
}
\section{Values}{

\itemize{
\item \code{1} benefits greater
\item \code{2} about equal (if volunteered)
\item \code{3} harmful results greater
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab B/C/- \tab partial \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/C/- \tab partial \cr
   2016 \tab A/-/- \tab partial \cr
   2018 \tab A/-/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3448/vshow}{GSS Data Explorer page for \code{nanowill}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{artseals}},
\code{\link{caremost}},
\code{\link{inuitway}},
\code{\link{nanoknw1}},
\code{\link{nanoknw2}},
\code{\link{nanotech}},
\code{\link{noicecap}},
\code{\link{penguins}},
\code{\link{sealevel}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
