% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_231.R
\name{natarts}
\alias{natarts}
\title{Spending right amount on arts}
\source{
General Social Survey https://gss.norc.org
}
\description{
natarts
}
\details{
Question Earlier we talked about problems in this country, and whether we are spending too much money on them, too little money, or about the right amount. I have one more problem to ask you about: the arts . . . are we spending too much, too little, or about the right amount on that?
}
\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2123/vshow}{GSS Data Explorer page for \code{natarts}}.
}

\seealso{
Other Religion and Culture: 
\code{\link{artgod}},
\code{\link{artists}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{relexp}},
\code{\link{relpersn}},
\code{\link{sprtprsn}}
}
\concept{Religion and Culture}
\keyword{variable}
