% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_338.R
\name{natborn}
\alias{natborn}
\title{What country r were born in}
\source{
General Social Survey https://gss.norc.org
}
\description{
natborn
}
\details{
Question What country were you born in?
}
\section{Values}{

\itemize{
\item \code{1} mexico
\item \code{2} canada
\item \code{3} china
\item \code{4} colombia
\item \code{5} cuba
\item \code{6} germany
\item \code{7} guatemala
\item \code{8} haiti
\item \code{9} honduras
\item \code{10} india
\item \code{11} jamaica
\item \code{12} korea
\item \code{13} philippines
\item \code{14} poland
\item \code{15} other
\item \code{16} trinidad
\item \code{17} antigua
\item \code{18} pakistan
\item \code{19} belize
\item \code{20} guyana
\item \code{21} egypt
\item \code{22} moldovia
\item \code{23} brazil
\item \code{24} portugal
\item \code{25} panama
\item \code{26} iran
\item \code{27} ghana
\item \code{28} bangladesh
\item \code{29} puerto rico
\item \code{30} turkey
\item \code{31} italy
\item \code{32} el salvidor
\item \code{33} greece
\item \code{34} ecuador
\item \code{35} england
\item \code{36} thailand
\item \code{37} peru
\item \code{38} australia
\item \code{39} israel
\item \code{40} netherlands
\item \code{41} macedonia
\item \code{42} france
\item \code{43} argentina
\item \code{44} lebanon
\item \code{45} russia
\item \code{46} cambodia
\item \code{47} venezuela
\item \code{48} iraq
\item \code{49} romania
\item \code{50} czech republic
\item \code{51} nicaragua
\item \code{52} arabic
\item \code{53} gambia
\item \code{54} west africa
\item \code{55} barbados
\item \code{56} south africa
\item \code{57} ukraine
\item \code{58} hungary
\item \code{59} eritrea
\item \code{60} ireland
\item \code{61} laos
\item \code{62} vietnam
\item \code{63} senegal
\item \code{64} sudan
\item \code{65} nigeria
\item \code{66} bosnia
\item \code{67} uruguay
\item \code{68} malaysia
\item \code{69} armenia
\item \code{70} congo
\item \code{71} japan
\item \code{72} dutch east indies
\item \code{73} indonesia
\item \code{74} africa
\item \code{75} jordan
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3024/vshow}{GSS Data Explorer page for \code{natborn}}.
}

\seealso{
Other Immigration: 
\code{\link{bornhome}},
\code{\link{immstats}},
\code{\link{robject}},
\code{\link{rundrstd}},
\code{\link{travelus}}
}
\concept{Immigration}
\keyword{variable}
