% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_426.R
\name{natdeg}
\alias{natdeg}
\title{In which country did you obtain your highest degree?}
\source{
General Social Survey https://gss.norc.org
}
\description{
natdeg
}
\details{
Question In which country did you obtain your highest degree?
}
\section{Values}{

\itemize{
\item \code{1} usa
\item \code{2} antigua
\item \code{3} trinidad
\item \code{4} jamaica
\item \code{5} pakistan
\item \code{6} turkey
\item \code{7} italy
\item \code{8} west indies
\item \code{9} colombia
\item \code{10} kazakhstan
\item \code{12} cuba
\item \code{13} india
\item \code{14} germany
\item \code{15} ecuador
\item \code{16} mexico
\item \code{17} portugal
\item \code{18} el salvador
\item \code{19} canada
\item \code{20} guatemala
\item \code{21} zimbabwe
\item \code{23} taiwan
\item \code{24} philippines
\item \code{25} israel
\item \code{26} dominican republic
\item \code{27} scotland
\item \code{28} haiti
\item \code{29} honduras
\item \code{30} puerto rico
\item \code{31} austraila
\item \code{32} china
\item \code{33} poland
\item \code{34} lithuania
\item \code{35} liberia
\item \code{36} brazil
\item \code{37} morocco
\item \code{38} spain
\item \code{39} bulgeria
\item \code{40} ussr
\item \code{41} kuwait
\item \code{42} japan
\item \code{43} korea
\item \code{44} vietnam
\item \code{45} bermuda
\item \code{46} ghana
\item \code{47} panama
\item \code{48} france
\item \code{49} romania
\item \code{50} south africa
\item \code{51} england
\item \code{52} britain
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3902/vshow}{GSS Data Explorer page for \code{natdeg}}.
}

\seealso{
Other Global Economics: 
\code{\link{agecmeus}},
\code{\link{bornsp}},
\code{\link{emphlth}},
\code{\link{emphplan}},
\code{\link{othplan}},
\code{\link{typevisa}},
\code{\link{visa}}

Other Immigration: 
\code{\link{agecmeus}},
\code{\link{bornsp}},
\code{\link{typevisa}},
\code{\link{visa}}
}
\concept{Education}
\concept{Global Economics}
\concept{Immigration}
\keyword{variable}
