% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_407.R
\name{natdiscm}
\alias{natdiscm}
\title{Should rebuilding community in disaster be govt resp}
\source{
General Social Survey https://gss.norc.org
}
\description{
natdiscm
}
\details{
Question B. Do you think it should or should not be the government's responsibility to help rebuild communities affected by natural disasters?
}
\section{Values}{

\itemize{
\item \code{1} should
\item \code{2} should not be
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3709/vshow}{GSS Data Explorer page for \code{natdiscm}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab should \tab should not be \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2992 \tab 1333 \tab 160 \tab 24 \tab 1 \tab 4510 \cr
   Total \tab 2992 \tab 1333 \tab 160 \tab 24 \tab 1 \tab 4510 \cr
}
}

\seealso{
Other Misc I: 
\code{\link{getaheay}},
\code{\link{newsprnt}}
}
\concept{Misc I}
\keyword{variable}
