% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_406.R
\name{natdisin}
\alias{natdisin}
\title{Should helping ppl in natural disaster be govt resp}
\source{
General Social Survey https://gss.norc.org
}
\description{
natdisin
}
\details{
Question A. Do you think it should or should not be the government's responsibility to help individuals affected by natural disasters?
}
\section{Values}{

\itemize{
\item \verb{[1]} should
\item \verb{[2]} should not be
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3708/vshow}{GSS Data Explorer page for \code{natdisin}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab should \tab should not be \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2992 \tab 1368 \tab 131 \tab 18 \tab 1 \tab 4510 \cr
   Total \tab 2992 \tab 1368 \tab 131 \tab 18 \tab 1 \tab 4510 \cr
}
}

\seealso{
Other Misc I: 
\code{\link{spnatdis}}
}
\concept{Misc I}
\keyword{variable}
