% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_406.R
\name{natdisin}
\alias{natdisin}
\title{Should helping ppl in natural disaster be govt resp}
\source{
General Social Survey https://gss.norc.org
}
\description{
natdisin
}
\details{
Question A. Do you think it should or should not be the government's responsibility to help individuals affected by natural disasters?
}
\section{Values}{

\itemize{
\item \code{1} should
\item \code{2} should not be
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3708/vshow}{GSS Data Explorer page for \code{natdisin}}.
}

\seealso{
Other Misc I: 
\code{\link{spnatdis}}

Other Federal Government: 
\code{\link{spnatdis}}
}
\concept{Federal Government}
\concept{Misc I}
\keyword{variable}
