% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_24.R
\name{natdrug}
\alias{natdrug}
\title{Dealing with drug addiction}
\source{
General Social Survey https://gss.norc.org
}
\description{
natdrug
}
\details{
Question We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. First (READ ITEM A) . . . are we spending too much, too little, or about the right amount on (ITEM)?
F. Dealing with drug addiction
}
\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/-/- \tab full \cr
   1974 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1976 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/186/vshow}{GSS Data Explorer page for \code{natdrug}}.
}

\seealso{
Other Core: 
\code{\link{nataid}},
\code{\link{natarms}},
\code{\link{natcity}},
\code{\link{natcrime}},
\code{\link{nateduc}},
\code{\link{natenvir}},
\code{\link{natheal}},
\code{\link{natrace}},
\code{\link{natspac}}

Other Federal Government: 
\code{\link{nataid}},
\code{\link{natarms}},
\code{\link{natcity}},
\code{\link{natcrime}},
\code{\link{nateduc}},
\code{\link{natenvir}},
\code{\link{natheal}},
\code{\link{natrace}},
\code{\link{natspac}}

Other Spending: 
\code{\link{nataid}},
\code{\link{natarms}},
\code{\link{natcity}},
\code{\link{natcrime}},
\code{\link{nateduc}},
\code{\link{natenvir}},
\code{\link{natheal}},
\code{\link{natrace}},
\code{\link{natspac}}

Other Split Ballots: 
\code{\link{nataid}},
\code{\link{natarms}},
\code{\link{natcity}},
\code{\link{natcrime}},
\code{\link{nateduc}},
\code{\link{natenvir}},
\code{\link{natheal}},
\code{\link{natrace}},
\code{\link{natspac}}

Other United States: 
\code{\link{nataid}},
\code{\link{natarms}},
\code{\link{natcity}},
\code{\link{natcrime}},
\code{\link{nateduc}},
\code{\link{natenvir}},
\code{\link{natheal}},
\code{\link{natrace}},
\code{\link{natspac}}
}
\concept{Core}
\concept{Drug Use And Addiction}
\concept{Federal Government}
\concept{Spending}
\concept{Split Ballots}
\concept{United States}
\keyword{variable}
