% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_24.R
\name{natheal}
\alias{natheal}
\title{Improving & protecting nations health}
\source{
General Social Survey https://gss.norc.org
}
\description{
natheal
}
\details{
Question We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. First (READ ITEM A) . . . are we spending too much, too little, or about the right amount on (ITEM)?
C. Improving and protecting the nation's health
}
\section{Values}{

\itemize{
\item \verb{[1]} too little
\item \verb{[2]} about right
\item \verb{[3]} too much
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   Full \tab 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/183/vshow}{GSS Data Explorer page for \code{natheal}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab too little \tab about right \tab too much \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1973 \tab - \tab 910 \tab 465 \tab 70 \tab 52 \tab 7 \tab - \tab 1504 \cr
   1974 \tab - \tab 944 \tab 415 \tab 67 \tab 51 \tab 7 \tab - \tab 1484 \cr
   1975 \tab - \tab 929 \tab 421 \tab 75 \tab 60 \tab 5 \tab - \tab 1490 \cr
   1976 \tab - \tab 902 \tab 465 \tab 74 \tab 50 \tab 8 \tab - \tab 1499 \cr
   1977 \tab - \tab 851 \tab 497 \tab 106 \tab 72 \tab 4 \tab - \tab 1530 \cr
   1978 \tab - \tab 849 \tab 516 \tab 106 \tab 61 \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 803 \tab 492 \tab 112 \tab 60 \tab 1 \tab - \tab 1468 \cr
   1982 \tab - \tab 1095 \tab 549 \tab 117 \tab 93 \tab 6 \tab - \tab 1860 \cr
   1983 \tab - \tab 913 \tab 536 \tab 82 \tab 60 \tab 8 \tab - \tab 1599 \cr
   1984 \tab 983 \tab 280 \tab 153 \tab 32 \tab 21 \tab 4 \tab - \tab 1473 \cr
   1985 \tab 783 \tab 432 \tab 248 \tab 47 \tab 22 \tab 2 \tab - \tab 1534 \cr
   1986 \tab 740 \tab 426 \tab 247 \tab 29 \tab 24 \tab 4 \tab - \tab 1470 \cr
   1987 \tab 1216 \tab 429 \tab 138 \tab 22 \tab 10 \tab 4 \tab - \tab 1819 \cr
   1988 \tab 763 \tab 471 \tab 198 \tab 23 \tab 22 \tab 4 \tab - \tab 1481 \cr
   1989 \tab 769 \tab 521 \tab 194 \tab 23 \tab 28 \tab 2 \tab - \tab 1537 \cr
   1990 \tab 698 \tab 484 \tab 151 \tab 18 \tab 17 \tab 4 \tab - \tab 1372 \cr
   1991 \tab 761 \tab 518 \tab 191 \tab 20 \tab 20 \tab 7 \tab - \tab 1517 \cr
   1993 \tab 807 \tab 571 \tab 137 \tab 60 \tab 26 \tab 5 \tab - \tab 1606 \cr
   1994 \tab 1484 \tab 964 \tab 349 \tab 127 \tab 62 \tab 6 \tab - \tab 2992 \cr
   1996 \tab 1463 \tab 950 \tab 331 \tab 113 \tab 42 \tab 5 \tab - \tab 2904 \cr
   1998 \tab 1451 \tab 922 \tab 345 \tab 78 \tab 34 \tab 2 \tab - \tab 2832 \cr
   2000 \tab 1408 \tab 1010 \tab 319 \tab 50 \tab 27 \tab 3 \tab - \tab 2817 \cr
   2002 \tab 1407 \tab 1000 \tab 284 \tab 52 \tab 21 \tab 1 \tab - \tab 2765 \cr
   2004 \tab 1411 \tab 1081 \tab 238 \tab 63 \tab 19 \tab - \tab - \tab 2812 \cr
   2006 \tab 3026 \tab 1094 \tab 282 \tab 75 \tab 32 \tab 1 \tab - \tab 4510 \cr
   2008 \tab 1019 \tab 749 \tab 177 \tab 47 \tab 31 \tab - \tab - \tab 2023 \cr
   2010 \tab 1030 \tab 597 \tab 234 \tab 157 \tab 25 \tab 1 \tab - \tab 2044 \cr
   2012 \tab 976 \tab 599 \tab 260 \tab 110 \tab 27 \tab 2 \tab - \tab 1974 \cr
   2014 \tab 1264 \tab 732 \tab 342 \tab 166 \tab 33 \tab 1 \tab - \tab 2538 \cr
   2016 \tab 1430 \tab 914 \tab 353 \tab 138 \tab 29 \tab 3 \tab - \tab 2867 \cr
   2018 \tab 1160 \tab 839 \tab 259 \tab 58 \tab 30 \tab 2 \tab - \tab 2348 \cr
   2021 \tab 2045 \tab 1314 \tab 483 \tab 169 \tab 5 \tab - \tab 16 \tab 4032 \cr
   2022 \tab 1764 \tab 1221 \tab 398 \tab 128 \tab 23 \tab 2 \tab 8 \tab 3544 \cr
   Total \tab 29858 \tab 26314 \tab 10667 \tab 2614 \tab 1189 \tab 111 \tab 24 \tab 70777 \cr
}
}

\seealso{
Other Core: 
\code{\link{nataid}},
\code{\link{natarms}},
\code{\link{natcity}},
\code{\link{natcrime}},
\code{\link{natdrug}},
\code{\link{nateduc}},
\code{\link{natenvir}},
\code{\link{natrace}},
\code{\link{natspac}}
}
\concept{Core}
\keyword{variable}
