% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_316.R
\name{natnotice}
\alias{natnotice}
\title{In my everyday life, i take notice of the natural elements around me (trees, wat}
\source{
General Social Survey https://gss.norc.org
}
\description{
natnotice
}
\details{
Question In my everyday life, I take notice of the natural elements around me (trees, water, wildlife, etc.).
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} somewhat agree
\item \verb{[3]} somewhat disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7184/vshow}{GSS Data Explorer page for \code{natnotice}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab somewhat agree \tab somewhat disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1175 \tab 815 \tab 303 \tab 20 \tab 8 \tab 6 \tab 21 \tab 2348 \cr
   Total \tab 1175 \tab 815 \tab 303 \tab 20 \tab 8 \tab 6 \tab 21 \tab 2348 \cr
}
}

\keyword{variable}
