% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_199.R
\name{natrecon}
\alias{natrecon}
\title{Natural environments should be left alone}
\source{
General Social Survey https://gss.norc.org
}
\description{
natrecon
}
\details{
Question How much do you agree or disagree with the following statements:
C. Natural environments that support scarce or endangered species should be left alone, no matter how great the economic benefits to your community from developing them commercially might be.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree somewhat
\item \verb{[3]} disagree somewhat
\item \verb{[4]} disagree strongly
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1873/vshow}{GSS Data Explorer page for \code{natrecon}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab agree strongly \tab agree somewhat \tab disagree somewhat \tab disagree strongly \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 441 \tab 374 \tab 290 \tab 245 \tab 82 \tab 12 \tab 2904 \cr
   Total \tab 1460 \tab 441 \tab 374 \tab 290 \tab 245 \tab 82 \tab 12 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{decauto}},
\code{\link{econsys}},
\code{\link{fambiz}},
\code{\link{famfinan}},
\code{\link{haggle}},
\code{\link{hrfambiz}},
\code{\link{reqinfo}},
\code{\link{tagsales}},
\code{\link{wkfambiz}}
}
\concept{Markets}
\keyword{variable}
