% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_28.R
\name{natroadz}
\alias{natroadz}
\title{Improving highways and bridges:ver z}
\source{
General Social Survey https://gss.norc.org
}
\description{
natroadz
}
\details{
Question We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. First (READ ITEM A) . . . are we spending too much, too little, or about the right amount on (ITEM)?
L. Improving the condition of highways and bridges
}
\section{Values}{

\itemize{
\item \verb{[1]} too little
\item \verb{[2]} about right
\item \verb{[3]} too much
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/221/vshow}{GSS Data Explorer page for \code{natroadz}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab too little \tab about right \tab too much \tab don't know \tab no answer \tab Total \cr
   1984 \tab 989 \tab 232 \tab 214 \tab 23 \tab 13 \tab 2 \tab 1473 \cr
   Total \tab 989 \tab 232 \tab 214 \tab 23 \tab 13 \tab 2 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{equal1}},
\code{\link{equal2}},
\code{\link{equal3}},
\code{\link{equal4}},
\code{\link{equal5}},
\code{\link{equal6}},
\code{\link{natmassz}},
\code{\link{natparkz}},
\code{\link{natsocz}}
}
\concept{Core}
\keyword{variable}
