% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_24.R
\name{natspac}
\alias{natspac}
\title{Space exploration program}
\source{
General Social Survey https://gss.norc.org
}
\description{
natspac
}
\details{
Question We are faced with many problems in this country, none of which can be solved easily or inexpensively. I'm going to name some of these problems, and for each one I'd like you to name some of these problems, and for each one I'd like you to tell me whether you think we're spending too much money on it, too little money, or about the right amount. First (READ ITEM A) . . . are we spending too much, too little, or about the right amount on (ITEM)?
A. Space exploration program
}
\section{Values}{

\itemize{
\item \code{1} too little
\item \code{2} about right
\item \code{3} too much
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/-/- \tab full \cr
   1974 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1976 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/181/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab about right \tab don't know \tab no answer \tab too little \tab too much \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 441 \tab 71 \tab 1 \tab 112 \tab 879 \tab - \tab 1504 \cr
   1974 \tab - \tab 408 \tab 53 \tab 4 \tab 114 \tab 905 \tab - \tab 1484 \cr
   1975 \tab - \tab 449 \tab 65 \tab - \tab 110 \tab 866 \tab - \tab 1490 \cr
   1976 \tab - \tab 419 \tab 37 \tab 3 \tab 137 \tab 903 \tab - \tab 1499 \cr
   1977 \tab - \tab 527 \tab 90 \tab - \tab 154 \tab 759 \tab - \tab 1530 \cr
   1978 \tab - \tab 536 \tab 96 \tab - \tab 177 \tab 723 \tab - \tab 1532 \cr
   1980 \tab - \tab 506 \tab 122 \tab 2 \tab 264 \tab 574 \tab - \tab 1468 \cr
   1982 \tab - \tab 682 \tab 121 \tab 2 \tab 201 \tab 854 \tab - \tab 1860 \cr
   1983 \tab - \tab 647 \tab 96 \tab 1 \tab 221 \tab 634 \tab - \tab 1599 \cr
   1984 \tab 983 \tab 210 \tab 27 \tab 3 \tab 58 \tab 192 \tab - \tab 1473 \cr
   1985 \tab 783 \tab 331 \tab 27 \tab 3 \tab 86 \tab 304 \tab - \tab 1534 \cr
   1986 \tab 740 \tab 316 \tab 30 \tab 3 \tab 82 \tab 299 \tab - \tab 1470 \cr
   1987 \tab 1216 \tab 204 \tab 38 \tab 1 \tab 87 \tab 273 \tab - \tab 1819 \cr
   1988 \tab 763 \tab 300 \tab 41 \tab 4 \tab 127 \tab 246 \tab - \tab 1481 \cr
   1989 \tab 769 \tab 338 \tab 49 \tab 2 \tab 113 \tab 266 \tab - \tab 1537 \cr
   1990 \tab 698 \tab 294 \tab 39 \tab 3 \tab 75 \tab 263 \tab - \tab 1372 \cr
   1991 \tab 761 \tab 321 \tab 56 \tab 6 \tab 88 \tab 285 \tab - \tab 1517 \cr
   1993 \tab 807 \tab 295 \tab 54 \tab 3 \tab 67 \tab 380 \tab - \tab 1606 \cr
   1994 \tab 1484 \tab 562 \tab 82 \tab 1 \tab 136 \tab 727 \tab - \tab 2992 \cr
   1996 \tab 1463 \tab 577 \tab 103 \tab 2 \tab 158 \tab 601 \tab - \tab 2904 \cr
   1998 \tab 1451 \tab 605 \tab 92 \tab 2 \tab 139 \tab 543 \tab - \tab 2832 \cr
   2000 \tab 1408 \tab 529 \tab 105 \tab 2 \tab 195 \tab 578 \tab - \tab 2817 \cr
   2002 \tab 1407 \tab 640 \tab 87 \tab 2 \tab 150 \tab 479 \tab - \tab 2765 \cr
   2004 \tab 1411 \tab 598 \tab 75 \tab 1 \tab 186 \tab 541 \tab - \tab 2812 \cr
   2006 \tab 3026 \tab 651 \tab 74 \tab 3 \tab 211 \tab 545 \tab - \tab 4510 \cr
   2008 \tab 1019 \tab 459 \tab 67 \tab 3 \tab 121 \tab 354 \tab - \tab 2023 \cr
   2010 \tab 1030 \tab 411 \tab 72 \tab 1 \tab 169 \tab 361 \tab - \tab 2044 \cr
   2012 \tab 976 \tab 399 \tab 94 \tab 5 \tab 203 \tab 297 \tab - \tab 1974 \cr
   2014 \tab 1264 \tab 518 \tab 121 \tab 3 \tab 298 \tab 334 \tab - \tab 2538 \cr
   2016 \tab 1430 \tab 661 \tab 144 \tab 3 \tab 281 \tab 348 \tab - \tab 2867 \cr
   2018 \tab 1160 \tab 528 \tab 158 \tab 1 \tab 230 \tab 271 \tab - \tab 2348 \cr
   2021 \tab 2045 \tab 1043 \tab 11 \tab - \tab 488 \tab 438 \tab 7 \tab 4032 \cr
   2022 \tab 1764 \tab 831 \tab 90 \tab 2 \tab 358 \tab 496 \tab 3 \tab 3544 \cr
   2024 \tab 1626 \tab 803 \tab 94 \tab 1 \tab 309 \tab 468 \tab 8 \tab 3309 \cr
   Total \tab 33097 \tab 17039 \tab 2581 \tab 73 \tab 5905 \tab 16986 \tab 18 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{nataid}},
\code{\link{natarms}},
\code{\link{natcity}},
\code{\link{natcrime}},
\code{\link{natdrug}},
\code{\link{nateduc}},
\code{\link{natenvir}},
\code{\link{natheal}},
\code{\link{natrace}}

Other Federal Government: 
\code{\link{nataid}},
\code{\link{natarms}},
\code{\link{natcity}},
\code{\link{natcrime}},
\code{\link{natdrug}},
\code{\link{nateduc}},
\code{\link{natenvir}},
\code{\link{natheal}},
\code{\link{natrace}}

Other Spending: 
\code{\link{nataid}},
\code{\link{natarms}},
\code{\link{natcity}},
\code{\link{natcrime}},
\code{\link{natdrug}},
\code{\link{nateduc}},
\code{\link{natenvir}},
\code{\link{natheal}},
\code{\link{natrace}}

Other Split Ballots: 
\code{\link{nataid}},
\code{\link{natarms}},
\code{\link{natcity}},
\code{\link{natcrime}},
\code{\link{natdrug}},
\code{\link{nateduc}},
\code{\link{natenvir}},
\code{\link{natheal}},
\code{\link{natrace}}

Other United States: 
\code{\link{nataid}},
\code{\link{natarms}},
\code{\link{natcity}},
\code{\link{natcrime}},
\code{\link{natdrug}},
\code{\link{nateduc}},
\code{\link{natenvir}},
\code{\link{natheal}},
\code{\link{natrace}}
}
\concept{Core}
\concept{Federal Government}
\concept{Spending}
\concept{Split Ballots}
\concept{United States}
\keyword{variable}
