% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_628.R
\name{naturdev}
\alias{naturdev}
\title{Accept a reduction in the size of protected nature areas in america}
\source{
General Social Survey https://gss.norc.org
}
\description{
naturdev
}
\details{
Question How willing would you be to accept a reduction in the size of America’s protected nature areas, in order to open them up for economic development?
}
\section{Values}{

\itemize{
\item \code{1} very willing
\item \code{2} fairly willing
\item \code{3} neither willing nor unwilling
\item \code{4} fairly unwilling
\item \code{5} very unwilling
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7489/vshow}{GSS Data Explorer page for \code{naturdev}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very willing \tab fairly willing \tab neither willing nor unwilling \tab fairly unwilling \tab very unwilling \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2003 \tab 28 \tab 159 \tab 303 \tab 513 \tab 768 \tab 71 \tab 176 \tab 11 \tab 4032 \cr
   Total \tab 2003 \tab 28 \tab 159 \tab 303 \tab 513 \tab 768 \tab 71 \tab 176 \tab 11 \tab 4032 \cr
}
}

\keyword{variable}
