% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_520.R
\name{naturgod}
\alias{naturgod}
\title{Relationship of nature to god}
\source{
General Social Survey https://gss.norc.org
}
\description{
naturgod
}
\details{
Question Please check one box to show which statement is closest to your views.
}
\section{Values}{

\itemize{
\item \verb{[1]} nature created by god
\item \verb{[2]} nature sacred in itself
\item \verb{[3]} nature import, but not sacred
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4735/vshow}{GSS Data Explorer page for \code{naturgod}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab nature created by god \tab nature sacred in itself \tab nature import, but not sacred \tab don't know \tab no answer \tab Total \cr
   1993 \tab 49 \tab 554 \tab 304 \tab 493 \tab 130 \tab 76 \tab 1606 \cr
   1994 \tab 1606 \tab 460 \tab 285 \tab 464 \tab 129 \tab 48 \tab 2992 \cr
   2000 \tab 1541 \tab 472 \tab 277 \tab 326 \tab 136 \tab 65 \tab 2817 \cr
   Total \tab 3196 \tab 1486 \tab 866 \tab 1283 \tab 395 \tab 189 \tab 7415 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{antests}},
\code{\link{grnprice}},
\code{\link{grnprog}},
\code{\link{grntaxes}},
\code{\link{grwtharm}},
\code{\link{grwthelp}},
\code{\link{naturpax}},
\code{\link{naturwar}},
\code{\link{resnatur}}
}
\concept{ISSP Environment}
\keyword{variable}
