% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_520.R
\name{naturpax}
\alias{naturpax}
\title{Nature would be at peace if humans left alone}
\source{
General Social Survey https://gss.norc.org
}
\description{
naturpax
}
\details{
Question How much do you agree or disagree with each of the following statements.
B. Nature would be at peace and in harmony if only human beings would leave it alone.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4730/vshow}{GSS Data Explorer page for \code{naturpax}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1993 \tab 49 \tab 106 \tab 625 \tab 317 \tab 390 \tab 40 \tab 23 \tab 56 \tab 1606 \cr
   1994 \tab 1606 \tab 118 \tab 554 \tab 283 \tab 331 \tab 49 \tab 24 \tab 27 \tab 2992 \cr
   Total \tab 1655 \tab 224 \tab 1179 \tab 600 \tab 721 \tab 89 \tab 47 \tab 83 \tab 4598 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{antests}},
\code{\link{grnprice}},
\code{\link{grnprog}},
\code{\link{grntaxes}},
\code{\link{grwtharm}},
\code{\link{grwthelp}},
\code{\link{naturgod}},
\code{\link{naturwar}},
\code{\link{resnatur}}
}
\concept{ISSP Environment}
\keyword{variable}
