% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_670.R
\name{neaeligible}
\alias{neaeligible}
\title{Nea follow-on eligible}
\source{
General Social Survey https://gss.norc.org
}
\description{
neaeligible
}
\details{
Question
}
\section{Values}{

\itemize{
\item \verb{[0]} not eligible for follow-on
\item \verb{[1]} eligible for follow-on
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8108/vshow}{GSS Data Explorer page for \code{neaeligible}}.

Counts by year:\tabular{lllll}{
   year \tab not available in this year \tab not eligible for follow-on \tab eligible for follow-on \tab Total \cr
   2022 \tab - \tab 1275 \tab 2269 \tab 3544 \cr
   Total \tab 0 \tab 1275 \tab 2269 \tab 3544 \cr
}
}

\concept{NEA}
\keyword{variable}
