% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_38.R
\name{neargod}
\alias{neargod}
\title{How close does r feel to god}
\source{
General Social Survey https://gss.norc.org
}
\description{
neargod
}
\details{
Question How close do you feel to God most of the time?  Would you say extremely close, somewhat close, not very close, or not close at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} extremely close
\item \verb{[2]} somewhat close
\item \verb{[3]} not very close
\item \verb{[4]} not close at all
\item \verb{[5]} does not believe
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1989 \cr
   A/B/C \tab 1988, 1991 \cr
   Full \tab 1983, 1984, 1985, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/321/vshow}{GSS Data Explorer page for \code{neargod}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab extremely close \tab somewhat close \tab not very close \tab not close at all \tab does not believe \tab don't know \tab no answer \tab Total \cr
   1983 \tab - \tab 471 \tab 838 \tab 151 \tab 91 \tab 20 \tab 17 \tab 11 \tab 1599 \cr
   1984 \tab - \tab 443 \tab 782 \tab 142 \tab 59 \tab 16 \tab 8 \tab 23 \tab 1473 \cr
   1985 \tab - \tab 495 \tab 758 \tab 156 \tab 81 \tab 14 \tab 17 \tab 13 \tab 1534 \cr
   1987 \tab - \tab 606 \tab 943 \tab 147 \tab 65 \tab 20 \tab 20 \tab 18 \tab 1819 \cr
   1988 \tab - \tab 428 \tab 775 \tab 142 \tab 84 \tab 24 \tab 22 \tab 6 \tab 1481 \cr
   1989 \tab 531 \tab 291 \tab 526 \tab 89 \tab 64 \tab 21 \tab 10 \tab 5 \tab 1537 \cr
   1991 \tab 158 \tab 449 \tab 616 \tab 121 \tab 56 \tab 25 \tab 66 \tab 26 \tab 1517 \cr
   Total \tab 689 \tab 3183 \tab 5238 \tab 948 \tab 500 \tab 140 \tab 160 \tab 102 \tab 10960 \cr
}
}

\seealso{
Other Core: 
\code{\link{creator}},
\code{\link{father}},
\code{\link{friend}},
\code{\link{judge}},
\code{\link{lover}},
\code{\link{master}},
\code{\link{mother}},
\code{\link{redeemer}},
\code{\link{spouse}}
}
\concept{Core}
\keyword{variable}
