% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_209.R
\name{needynei}
\alias{needynei}
\title{Contributed to a needy neighbor.}
\source{
General Social Survey https://gss.norc.org
}
\description{
needynei
}
\details{
Question People help other people in ways besides giving money, time, or other things to organized groups. Sometimes people help needy people directly. During the past 12 months, did you or members of your family or household give money, food, or clothing to any of the following types of people?
B. A needy neighbor
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1976/vshow}{GSS Data Explorer page for \code{needynei}}.
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{homeless}},
\code{\link{needyfrd}},
\code{\link{needyoth}},
\code{\link{needyrel}},
\code{\link{valinfrm}},
\code{\link{valintl}},
\code{\link{valoth}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
