% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_428.R
\name{negjob5}
\alias{negjob5}
\title{In last 5 yrs, r been denied promo or recvd bad evaluation bc employr belivd r w}
\source{
General Social Survey https://gss.norc.org
}
\description{
negjob5
}
\details{
Question In the last five years, have you been denied a promotion or received a negative job evaluation because an employer believed you were gay, lesbian, or bisexual?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} not employed in last 5 years
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3923/vshow}{GSS Data Explorer page for \code{negjob5}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab not employed in last 5 years \tab Total \cr
   2008 \tab 1989 \tab 3 \tab 1 \tab 30 \tab - \tab 2023 \cr
   2012 \tab 1918 \tab 5 \tab 1 \tab 47 \tab 3 \tab 1974 \cr
   Total \tab 3907 \tab 8 \tab 2 \tab 77 \tab 3 \tab 3997 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{dwell5}},
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{kidnow2}},
\code{\link{losejob5}},
\code{\link{toldwork}}
}
\concept{Sexual Orientation}
\keyword{variable}
