% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_458.R
\name{neifrd}
\alias{neifrd}
\title{Number of close friends who live near r}
\source{
General Social Survey https://gss.norc.org
}
\description{
neifrd
}
\details{
Question Thinking now of people who live near you - in your neighborhood or district: How many of these people are close friends of yours?
}
\section{Values}{

\itemize{
\item \code{0} no close friends living near r
\item \code{96} more than a hundred
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4203/vshow}{GSS Data Explorer page for \code{neifrd}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{bestfrd}},
\code{\link{bstcall}},
\code{\link{bstvisit}},
\code{\link{cowrkfrd}},
\code{\link{godparts}},
\code{\link{grpchurh}},
\code{\link{grppol}},
\code{\link{grpsport}},
\code{\link{grpunion}}

Other Friends: 
\code{\link{bestfrd}},
\code{\link{bstcall}},
\code{\link{bstvisit}},
\code{\link{cowrkfrd}}
}
\concept{Friends}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
