% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_611.R
\name{neisafe}
\alias{neisafe}
\title{How safe interviewer thinks neighborhood is}
\source{
General Social Survey https://gss.norc.org
}
\description{
neisafe
}
\details{
Question Now thinking about safety in this neighborhood, how safe do you think it is?
}
\section{Values}{

\itemize{
\item \code{1} very safe
\item \code{2} somewhat safe
\item \code{3} somewhat unsafe
\item \code{4} very unsafe
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6658/vshow}{GSS Data Explorer page for \code{neisafe}}.
}

\keyword{variable}
