% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_155.R
\name{neverwk}
\alias{neverwk}
\title{R never worked full-time after school}
\source{
General Social Survey https://gss.norc.org
}
\description{
neverwk
}
\details{
Question RESPONDENT'S WORK HISTORY
}
\section{Values}{

\itemize{
\item \code{1} never worked
\item \code{2} worked
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1461/vshow}{GSS Data Explorer page for \code{neverwk}}.
}

\seealso{
Other Family Mobility: 
\code{\link{alike1}},
\code{\link{alike2}},
\code{\link{datefrst}},
\code{\link{datesch}},
\code{\link{indfirst}},
\code{\link{occfirst}},
\code{\link{payrdied}},
\code{\link{presfrst}},
\code{\link{wrkslfst}}
}
\concept{Family Mobility}
\keyword{variable}
