% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_507.R
\name{newjob}
\alias{newjob}
\title{How helpful present work skill would be for new job}
\source{
General Social Survey https://gss.norc.org
}
\description{
newjob
}
\details{
Question If you were to look for a new job, how helpful would your present work experience and/or job skills be?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} likely
\item \verb{[3]} unlikely
\item \verb{[4]} very unlikely
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4616/vshow}{GSS Data Explorer page for \code{newjob}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab likely \tab unlikely \tab very unlikely \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 3492 \tab 567 \tab 347 \tab 78 \tab 22 \tab 2 \tab 2 \tab 4510 \cr
   Total \tab 3492 \tab 567 \tab 347 \tab 78 \tab 22 \tab 2 \tab 2 \tab 4510 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{dailywrk}},
\code{\link{famintjb}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobeasy}},
\code{\link{nounemp1}},
\code{\link{nounemp2}},
\code{\link{replaceu}},
\code{\link{timeoff}}
}
\concept{Work Orientation}
\keyword{variable}
