% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_267.R
\name{newjob12}
\alias{newjob12}
\title{R use web to look for new job in past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
newjob12
}
\details{
Question In the past 12 months, that is, from (INSERT CURRENT MONTH) 1999/2001 through (INSERT CURRENT MONTH) 2000/02 have you used the Web to do each of these things:
K. Look for a new job or explore career opportunities?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2425/vshow}{GSS Data Explorer page for \code{newjob12}}.
}

\seealso{
Other Information Society: 
\code{\link{abort12}},
\code{\link{chat12}},
\code{\link{econ12}},
\code{\link{game12}},
\code{\link{health12}},
\code{\link{locate12}},
\code{\link{people12}},
\code{\link{polinf12}},
\code{\link{travel12}}

Other Computers And Internet: 
\code{\link{abort12}},
\code{\link{chat12}},
\code{\link{econ12}},
\code{\link{game12}},
\code{\link{health12}},
\code{\link{locate12}},
\code{\link{people12}},
\code{\link{polinf12}},
\code{\link{travel12}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
