% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_424.R
\name{newjobs}
\alias{newjobs}
\title{Jobs created are better than r's current job}
\source{
General Social Survey https://gss.norc.org
}
\description{
newjobs
}
\details{
Question On average, do you think the jobs that have been created in the United States for people with jobs like yours are better than your job (with higher pay and better benefits) or worse?
}
\section{Values}{

\itemize{
\item \code{1} better
\item \code{2} about the same
\item \code{3} worse
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3885/vshow}{GSS Data Explorer page for \code{newjobs}}.
}

\seealso{
Other Global Economics: 
\code{\link{findnwjb}},
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{moretrde}},
\code{\link{whyjbct2}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{wkcomptr}}

Other Employment: 
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{moretrde}},
\code{\link{whyjbct2}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}}

Other Job: 
\code{\link{findnwjb}},
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{moretrde}},
\code{\link{whyjbct2}},
\code{\link{wkcomptr}}

Other Unemployment: 
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{moretrde}},
\code{\link{whyjbct2}}
}
\concept{Employment}
\concept{Global Economics}
\concept{Job}
\concept{Technology}
\concept{Unemployment}
\concept{Work}
\keyword{variable}
