% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_386.R
\name{newsfrmy}
\alias{newsfrmy}
\title{Where news events info}
\source{
General Social Survey https://gss.norc.org
}
\description{
newsfrmy
}
\details{
Question We are interested in how people get information about events in the news. Where do you get most of your information about current news events – printed newspapers,  online newspapers, printed magazines, online magazines, other places on the Internet, books or other printed materials, TV, radio, family, friends, colleagues, or some other source?
}
\section{Values}{

\itemize{
\item \code{1} print newspapers
\item \code{2} online newspapers
\item \code{3} print magazines
\item \code{4} online magazines
\item \code{5} other places on the internet
\item \code{6} books/other printed material
\item \code{7} tv
\item \code{8} radio
\item \code{9} family
\item \code{10} friends/colleagues
\item \code{11} other
\item \code{98} don~~t know
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3503/vshow}{GSS Data Explorer page for \code{newsfrmy}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab print newspapers \tab online newspapers \tab print magazines \tab online magazines \tab other places on the internet \tab books/other printed material \tab tv \tab radio \tab family \tab friends/colleagues \tab other \tab no answer \tab Total \cr
   2010 \tab 1386 \tab 97 \tab 83 \tab 6 \tab 2 \tab 57 \tab 5 \tab 328 \tab 40 \tab 16 \tab 16 \tab 6 \tab 2 \tab 2044 \cr
   Total \tab 1386 \tab 97 \tab 83 \tab 6 \tab 2 \tab 57 \tab 5 \tab 328 \tab 40 \tab 16 \tab 16 \tab 6 \tab 2 \tab 2044 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{cloning}},
\code{\link{scientda}},
\code{\link{scientsn}},
\code{\link{scifromy}},
\code{\link{sciinfgo}},
\code{\link{scinews1}},
\code{\link{scinews2}},
\code{\link{scinews3}},
\code{\link{seeksciy}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
