% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_291.R
\name{newsgrps}
\alias{newsgrps}
\title{How familiar r is with term newsgroups}
\source{
General Social Survey https://gss.norc.org
}
\description{
newsgrps
}
\details{
Question I. Newsgroups
}
\section{Values}{

\itemize{
\item \verb{[1]} very familiar
\item \verb{[2]} somewhat familiar
\item \verb{[3]} not familiar
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2638/vshow}{GSS Data Explorer page for \code{newsgrps}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very familiar \tab somewhat familiar \tab not familiar \tab don't know \tab no answer \tab Total \cr
   2002 \tab 2142 \tab 158 \tab 187 \tab 230 \tab 1 \tab 47 \tab 2765 \cr
   2004 \tab 992 \tab 510 \tab 600 \tab 689 \tab 1 \tab 20 \tab 2812 \cr
   Total \tab 3134 \tab 668 \tab 787 \tab 919 \tab 2 \tab 67 \tab 5577 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{advsrch}},
\code{\link{browser3}},
\code{\link{ezines}},
\code{\link{mp3}},
\code{\link{prefsets}},
\code{\link{timekid1}},
\code{\link{timekid2}},
\code{\link{timekid3}},
\code{\link{timekid4}}
}
\concept{Information Society}
\keyword{variable}
