% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_407.R
\name{newsprnt}
\alias{newsprnt}
\title{How often does r read traditional newspaper}
\source{
General Social Survey https://gss.norc.org
}
\description{
newsprnt
}
\details{
Question How often do you read a traditional newspaper (i.e. one that is printed with ink on paper and NOT counting what you read online) -- every day, a few times a week, once a week, less than once a week, or never?
}
\section{Values}{

\itemize{
\item \code{1} every day
\item \code{2} a few times a week
\item \code{3} once a week
\item \code{4} less than once a week
\item \code{5} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3710/vshow}{GSS Data Explorer page for \code{newsprnt}}.
}

\seealso{
Other Misc I: 
\code{\link{getaheay}},
\code{\link{natdiscm}},
\code{\link{natdisin}},
\code{\link{spnatdis}}
}
\concept{Misc I}
\keyword{variable}
