% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_191.R
\name{newused}
\alias{newused}
\title{Was the car new or used?}
\source{
General Social Survey https://gss.norc.org
}
\description{
newused
}
\details{
Question Was the car new or used?
}
\section{Values}{

\itemize{
\item \verb{[1]} new
\item \verb{[2]} used
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1800/vshow}{GSS Data Explorer page for \code{newused}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab new \tab used \tab no answer \tab Total \cr
   1996 \tab 2497 \tab 209 \tab 191 \tab 7 \tab 2904 \cr
   Total \tab 2497 \tab 209 \tab 191 \tab 7 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{cardealr}},
\code{\link{carprivt}},
\code{\link{relprivt}},
\code{\link{typdealr}},
\code{\link{warrntyc}}
}
\concept{Markets}
\keyword{variable}
