% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_663.R
\name{nextstatus}
\alias{nextstatus}
\title{Gss next follow-on status}
\source{
General Social Survey https://gss.norc.org
}
\description{
nextstatus
}
\details{
Question FLAG INDICATING WHETHER OR NOT THE RESPONDENT PARTICIPATED IN THE GSS NEXT FOLLOW-ON
}
\section{Values}{

\itemize{
\item \code{1} completed
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7962/vshow}{GSS Data Explorer page for \code{nextstatus}}.
}

\keyword{variable}
