% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_457.R
\name{niecenep}
\alias{niecenep}
\title{How often does r contact nieces and nephews}
\source{
General Social Survey https://gss.norc.org
}
\description{
niecenep
}
\details{
Question Now some questions about your contact with other relatives. Please indicate how often you have been in contact with any of the following types of relatives in the last four weeks.
E. Nieces and nephews
}
\section{Values}{

\itemize{
\item \code{1} more than twice in last 4 weeks
\item \code{2} once or twice in last 4 weeks
\item \code{3} not at all in last 4 weeks
\item \code{4} i have no living relative of this type
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4200/vshow}{GSS Data Explorer page for \code{niecenep}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{cousins}},
\code{\link{kidcall}},
\code{\link{macall1}},
\code{\link{mavisit1}},
\code{\link{pacall1}},
\code{\link{parslaw}},
\code{\link{pavisit1}},
\code{\link{sibinlaw}},
\code{\link{uncaunts}}

Other Family: 
\code{\link{cousins}},
\code{\link{kidcall}},
\code{\link{macall1}},
\code{\link{mavisit1}},
\code{\link{pacall1}},
\code{\link{parslaw}},
\code{\link{pavisit1}},
\code{\link{sibinlaw}},
\code{\link{uncaunts}}
}
\concept{Family}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
