% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_516.R
\name{nirvana}
\alias{nirvana}
\title{Believe in nirvana}
\source{
General Social Survey https://gss.norc.org
}
\description{
nirvana
}
\details{
Question Do you believe in…
B. ...in Nirvana?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, definitely
\item \verb{[2]} yes, probably
\item \verb{[3]} no, probably not
\item \verb{[4]} no, definitely not
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4706/vshow}{GSS Data Explorer page for \code{nirvana}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes, definitely \tab yes, probably \tab no, probably not \tab no, definitely not \tab don't know \tab no answer \tab Total \cr
   2008 \tab 658 \tab 61 \tab 194 \tab 312 \tab 346 \tab 434 \tab 18 \tab 2023 \cr
   Total \tab 658 \tab 61 \tab 194 \tab 312 \tab 346 \tab 434 \tab 18 \tab 2023 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{ancestrs}},
\code{\link{mywaygod}},
\code{\link{reincar}},
\code{\link{relcand}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relmarry}}
}
\concept{ISSP Religion}
\keyword{variable}
