% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_528.R
\name{nobuygrn}
\alias{nobuygrn}
\title{How often avoid products}
\source{
General Social Survey https://gss.norc.org
}
\description{
nobuygrn
}
\details{
Question K3. And how often do you avoid buying certain products for environmental reasons?
}
\section{Values}{

\itemize{
\item \code{1} always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} never
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2010, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4815/vshow}{GSS Data Explorer page for \code{nobuygrn}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab always \tab often \tab sometimes \tab never \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2010 \tab 614 \tab 105 \tab 351 \tab 540 \tab 411 \tab 16 \tab 7 \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 146 \tab 525 \tab 769 \tab 381 \tab 2 \tab 194 \tab 12 \tab 4032 \cr
   Total \tab 2617 \tab 251 \tab 876 \tab 1309 \tab 792 \tab 18 \tab 201 \tab 12 \tab 6076 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{busgrn}},
\code{\link{grneffme}},
\code{\link{h2oless}},
\code{\link{helpharm}},
\code{\link{peopgrn}},
\code{\link{redcehme}},
\code{\link{tempgen1}}
}
\concept{ISSP Environment}
\keyword{variable}
