% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_683.R
\name{nocondom}
\alias{nocondom}
\title{Had vaginal or anal sex without a condom in past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
nocondom
}
\details{
Question Did you, or did you not, have vaginal or anal sex without a condom in the past 12 months?
}
\section{Values}{

\itemize{
\item \code{1} i had vaginal or anal sex without a condom in the past 12 months
\item \code{2} i did not have vaginal or anal sex without a condom in the past 12 months
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8151/vshow}{GSS Data Explorer page for \code{nocondom}}.
}

\seealso{
Other High Risk Behaviors: 
\code{\link{nodocidu}}
}
\concept{High Risk Behaviors}
\keyword{variable}
