% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_683.R
\name{nodocidu}
\alias{nodocidu}
\title{Ever injected any drugs other than those prescribed for you}
\source{
General Social Survey https://gss.norc.org
}
\description{
nodocidu
}
\details{
Question Have you ever in your life injected any drugs other than those prescribed for you? Do not include anything you took under a doctor’s orders.
As of 2024, NODOCIDU has been slightly adjusted to help protect confidentiality. Please see the 2024 codebook for more information.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8200/vshow}{GSS Data Explorer page for \code{nodocidu}}.
}

\seealso{
Other High Risk Behaviors: 
\code{\link{nocondom}}
}
\concept{Health}
\concept{High Risk Behaviors}
\keyword{variable}
