% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_236.R
\name{nogo}
\alias{nogo}
\title{Perf or exhibit r wanted to go to in past 12 months but did not}
\source{
General Social Survey https://gss.norc.org
}
\description{
nogo
}
\details{
Question During the last 12 months, was there a performance or exhibit that you wanted to go to but did not?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6247/vshow}{GSS Data Explorer page for \code{nogo}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1390 \tab 616 \tab 851 \tab 2 \tab 8 \tab 2867 \cr
   Total \tab 1390 \tab 616 \tab 851 \tab 2 \tab 8 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{artwhy3}},
\code{\link{artwhy4}},
\code{\link{artwhy5}},
\code{\link{artwhy6}},
\code{\link{artwhy7}},
\code{\link{artwhy8}},
\code{\link{artwhy9}},
\code{\link{nogo1}},
\code{\link{prfmcost}}
}
\concept{Arts and Culture}
\keyword{variable}
