% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_294.R
\name{nogovt}
\alias{nogovt}
\title{No govt interference important for freedom}
\source{
General Social Survey https://gss.norc.org
}
\description{
nogovt
}
\details{
Question I’m going to read you several statements about freedom. Each one may say something true and important about freedom, but I’d like to know how important each statement is to you. Is it one of the most important things about freedom, extremely important, very important, somewhat important, or not too important?
B. Freedom is having a government that doesn’t spy on me or interfere in my life.
}
\section{Values}{

\itemize{
\item \code{1} one of the most important
\item \code{2} extremely important
\item \code{3} very important
\item \code{4} moderately important
\item \code{5} somewhat important
\item \code{6} not too important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2669/vshow}{GSS Data Explorer page for \code{nogovt}}.
}

\seealso{
Other Freedom: 
\code{\link{choice}},
\code{\link{cntrlife}},
\code{\link{expunpop}},
\code{\link{freeexp1}},
\code{\link{freeexp2}},
\code{\link{freeexp3}},
\code{\link{freemns1}},
\code{\link{freemns2}},
\code{\link{freemns3}},
\code{\link{freenow}},
\code{\link{freeprss}},
\code{\link{howfree}},
\code{\link{inpeace}},
\code{\link{leftlone}},
\code{\link{partpol}},
\code{\link{rfreenow}},
\code{\link{rhowfree}},
\code{\link{satdemoc}},
\code{\link{wlthpov}}
}
\concept{Freedom}
\keyword{variable}
