% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_653.R
\name{nointerest}
\alias{nointerest}
\title{How often felt no interest in doing things in past 2 weeks}
\source{
General Social Survey https://gss.norc.org
}
\description{
nointerest
}
\details{
Question Over the last 2 weeks, how often have you been bothered by the following problems: Little interest or pleasure in doing things.
}
\section{Values}{

\itemize{
\item \code{1} not at all
\item \code{2} several days
\item \code{3} more than half the days
\item \code{4} nearly every day
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7862/vshow}{GSS Data Explorer page for \code{nointerest}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab not at all \tab several days \tab more than half the days \tab nearly every day \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 1557 \tab 1329 \tab 447 \tab 100 \tab 66 \tab 6 \tab 19 \tab 20 \tab 3544 \cr
   Total \tab 1557 \tab 1329 \tab 447 \tab 100 \tab 66 \tab 6 \tab 19 \tab 20 \tab 3544 \cr
}
}

\keyword{variable}
