% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_524.R
\name{nomeat}
\alias{nomeat}
\title{Refuse to eat meat for envir reasons}
\source{
General Social Survey https://gss.norc.org
}
\description{
nomeat
}
\details{
Question C. And how often do you refuse to eat meat for moral or environmental reasons?
}
\section{Values}{

\itemize{
\item \code{1} always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} never
\item \code{5} no car, dont drive
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4771/vshow}{GSS Data Explorer page for \code{nomeat}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{chemfree}},
\code{\link{comtype}},
\code{\link{drivless}},
\code{\link{grndemo}},
\code{\link{grngroup}},
\code{\link{grnmoney}},
\code{\link{grnsign}},
\code{\link{recycle}},
\code{\link{usdoenuf}}
}
\concept{ISSP Environment}
\keyword{variable}
