% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_632.R
\name{norelgsp}
\alias{norelgsp}
\title{Is r atheist or agnostic}
\source{
General Social Survey https://gss.norc.org
}
\description{
norelgsp
}
\details{
Question Do you consider yourself to be atheist, agnostic, something else, or nothing in particular?
}
\section{Values}{

\itemize{
\item \verb{[1]} atheist
\item \verb{[2]} agnostic
\item \verb{[3]} something else
\item \verb{[4]} nothing in particular
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7522/vshow}{GSS Data Explorer page for \code{norelgsp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab atheist \tab agnostic \tab something else \tab nothing in particular \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2911 \tab 186 \tab 225 \tab 125 \tab 482 \tab 96 \tab 7 \tab 4032 \cr
   Total \tab 2911 \tab 186 \tab 225 \tab 125 \tab 482 \tab 96 \tab 7 \tab 4032 \cr
}
}

\keyword{variable}
