% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_632.R
\name{norelgsp}
\alias{norelgsp}
\title{Is r atheist or agnostic}
\source{
General Social Survey https://gss.norc.org
}
\description{
norelgsp
}
\details{
Question Do you consider yourself to be atheist, agnostic, something else, or nothing in particular?
}
\section{Values}{

\itemize{
\item \code{1} atheist
\item \code{2} agnostic
\item \code{3} something else
\item \code{4} nothing in particular
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7522/vshow}{GSS Data Explorer page for \code{norelgsp}}.
}

\keyword{variable}
