% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_190.R
\name{notam}
\alias{notam}
\title{If not american, how much change}
\source{
General Social Survey https://gss.norc.org
}
\description{
notam
}
\details{
Question If you were no longer an AMERICAN, but were a citizen of another country living outside the United States, how much would it change what you Do in a typical week?  Would what you do in a typical week be:
}
\section{Values}{

\itemize{
\item \code{1} completely different
\item \code{2} quite different
\item \code{3} somewhat different
\item \code{4} just slightly different
\item \code{5} not at all different
\item \code{7} citizen of another country
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1788/vshow}{GSS Data Explorer page for \code{notam}}.
}

\seealso{
Other Emotions: 
\code{\link{amimp}},
\code{\link{chnang16}},
\code{\link{likeoth}},
\code{\link{othresp}},
\code{\link{reactok}},
\code{\link{selfresp}},
\code{\link{usfrustr}},
\code{\link{ussat}},
\code{\link{usworry}}
}
\concept{Emotions}
\keyword{variable}
