% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_544.R
\name{notvote}
\alias{notvote}
\title{Citizens have right not to vote}
\source{
General Social Survey https://gss.norc.org
}
\description{
notvote
}
\details{
Question On a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it:
D. That citizens have the right not to vote.
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all important
\item \verb{[7]} very important
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4969/vshow}{GSS Data Explorer page for \code{notvote}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab not at all important \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab can't choose \tab no answer \tab Total \cr
   2014 \tab 1274 \tab 90 \tab 26 \tab 31 \tab 139 \tab 124 \tab 174 \tab 629 \tab 47 \tab 4 \tab 2538 \cr
   Total \tab 1274 \tab 90 \tab 26 \tab 31 \tab 139 \tab 124 \tab 174 \tab 629 \tab 47 \tab 4 \tab 2538 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{hlthall}},
\code{\link{leftrght}},
\code{\link{ntcitvte}}
}
\concept{ISSP Citizenship}
\keyword{variable}
