% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_544.R
\name{notvote}
\alias{notvote}
\title{Citizens have right not to vote}
\source{
General Social Survey https://gss.norc.org
}
\description{
notvote
}
\details{
Question On a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it:
D. That citizens have the right not to vote.
}
\section{Values}{

\itemize{
\item \code{1} not at all important
\item \code{7} very important
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4969/vshow}{GSS Data Explorer page for \code{notvote}}.
}

\seealso{
Other ISSP Citizenship: 
\code{\link{hlthall}},
\code{\link{leftrght}},
\code{\link{ntcitvte}}
}
\concept{ISSP Citizenship}
\keyword{variable}
