% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_507.R
\name{nounemp1}
\alias{nounemp1}
\title{R would accept job requiring new skill to avoid unemployment}
\source{
General Social Survey https://gss.norc.org
}
\description{
nounemp1
}
\details{
Question To what extent do you agree or disagree with the following statements? In order to avoid unemployment I would be willing…
A. To accept a job that requires new skills
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D/- \tab partial \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4620/vshow}{GSS Data Explorer page for \code{nounemp1}}.
}

\seealso{
Other Work Orientation: 
\code{\link{dailywrk}},
\code{\link{famintjb}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobeasy}},
\code{\link{newjob}},
\code{\link{nounemp2}},
\code{\link{replaceu}},
\code{\link{timeoff}}

Other Unemployment: 
\code{\link{nounemp2}}
}
\concept{Unemployment}
\concept{Work Orientation}
\keyword{variable}
