% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_508.R
\name{nounemp3}
\alias{nounemp3}
\title{R would accept temporary employment to avoid unemployment}
\source{
General Social Survey https://gss.norc.org
}
\description{
nounemp3
}
\details{
Question To what extent do you agree or disagree with the following statements? In order to avoid unemployment I would be willing…
C. To accept temporary employment
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D/- \tab partial \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4622/vshow}{GSS Data Explorer page for \code{nounemp3}}.
}

\seealso{
Other Work Orientation: 
\code{\link{nounemp4}},
\code{\link{otherwrk}},
\code{\link{skltrain}},
\code{\link{spwrkgvt}}
}
\concept{Work Orientation}
\keyword{variable}
