% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_672.R
\name{nouseint5}
\alias{nouseint5}
\title{No internet use over past 12 months due to lack of utility/importance}
\source{
General Social Survey https://gss.norc.org
}
\description{
nouseint5
}
\details{
Question Which of the following are reasons you have not used the Internet in the last 12 months?… The Internet is not important or useful for me.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8212/vshow}{GSS Data Explorer page for \code{nouseint5}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{intrnetuse}},
\code{\link{nouseint1}},
\code{\link{nouseint2}},
\code{\link{nouseint3}},
\code{\link{nouseint4}},
\code{\link{nouseint6}},
\code{\link{nouseint7}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
