% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_526.R
\name{nukeacc}
\alias{nukeacc}
\title{Likelihood of nuclear meltdown in 5 years}
\source{
General Social Survey https://gss.norc.org
}
\description{
nukeacc
}
\details{
Question Within the next five years, how likely is it that an accident at a nuclear power station will cause long-term environmental damage across many countries?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} likely
\item \verb{[3]} unlikely
\item \verb{[4]} very unlikely
\item \verb{[5]} cant choose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4790/vshow}{GSS Data Explorer page for \code{nukeacc}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab likely \tab unlikely \tab very unlikely \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1541 \tab 184 \tab 415 \tab 330 \tab 140 \tab 148 \tab 59 \tab 2817 \cr
   Total \tab 1541 \tab 184 \tab 415 \tab 330 \tab 140 \tab 148 \tab 59 \tab 2817 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{econgrn}},
\code{\link{excldimm}},
\code{\link{infobiz}},
\code{\link{infocol}},
\code{\link{infogovt}},
\code{\link{infogrn}},
\code{\link{infonews}},
\code{\link{infotv}},
\code{\link{ldcgrn}}
}
\concept{ISSP Environment}
\keyword{variable}
