% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_385.R
\name{nukeelec}
\alias{nukeelec}
\title{View about nuclear energy}
\source{
General Social Survey https://gss.norc.org
}
\description{
nukeelec
}
\details{
Question Please look at Card E16. When making policy recommendations about nuclear power, to what extent do you think the following groups would support what is best for the country as a whole or what serves their own narrow interests? FI INSTRUCTION: IF ASKED WHAT NARROW INTEREST MEANS: Well, someone might gain financially if a certain policy were adopted or it might advance his or her career.
D. Which statement best describes your own view about increasing the use of nuclear power to generate electricity in the U.S.?
}
\section{Values}{

\itemize{
\item \code{1} strongly favor
\item \code{2} favor
\item \code{3} oppose
\item \code{4} strongly oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3498/vshow}{GSS Data Explorer page for \code{nukeelec}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bizbstnk}},
\code{\link{bizinfnk}},
\code{\link{biznuke}},
\code{\link{engagrnk}},
\code{\link{engbstnk}},
\code{\link{engnuke}},
\code{\link{enhinfnk}},
\code{\link{polbstnk}},
\code{\link{polinfnk}}

Other Environment: 
\code{\link{bizbstnk}},
\code{\link{bizinfnk}},
\code{\link{biznuke}},
\code{\link{engagrnk}},
\code{\link{engbstnk}},
\code{\link{engnuke}},
\code{\link{enhinfnk}},
\code{\link{polbstnk}},
\code{\link{polinfnk}}
}
\concept{Environment}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
