% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_522.R
\name{nukefam}
\alias{nukefam}
\title{Nuke power danger to my family}
\source{
General Social Survey https://gss.norc.org
}
\description{
nukefam
}
\details{
Question B. And do you think that nuclear power stations are…
}
\section{Values}{

\itemize{
\item \code{1} extremely dangerous
\item \code{2} very dangerous
\item \code{3} somewhat dangerous
\item \code{4} not very dangerous
\item \code{5} not dangerous at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4757/vshow}{GSS Data Explorer page for \code{nukefam}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab extremely dangerous \tab very dangerous \tab somewhat dangerous \tab not very dangerous \tab not dangerous at all \tab can't choose \tab no answer \tab Total \cr
   1993 \tab 49 \tab 253 \tab 288 \tab 515 \tab 278 \tab 65 \tab 68 \tab 90 \tab 1606 \cr
   1994 \tab 1606 \tab 234 \tab 288 \tab 456 \tab 222 \tab 42 \tab 72 \tab 72 \tab 2992 \cr
   Total \tab 1655 \tab 487 \tab 576 \tab 971 \tab 500 \tab 107 \tab 140 \tab 162 \tab 4598 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{carsfam}},
\code{\link{carsgen}},
\code{\link{carsten}},
\code{\link{grntest3}},
\code{\link{grntest4}},
\code{\link{grntest5}},
\code{\link{grntest6}},
\code{\link{grntest7}},
\code{\link{nukegen}}
}
\concept{ISSP Environment}
\keyword{variable}
