% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_522.R
\name{nukegen}
\alias{nukegen}
\title{Nuke power danger to envir}
\source{
General Social Survey https://gss.norc.org
}
\description{
nukegen
}
\details{
Question A. In general, do you think that nuclear power stations are…
}
\section{Values}{

\itemize{
\item \code{1} extremely dangerous
\item \code{2} very dangerous
\item \code{3} somewhat dangerous
\item \code{4} not very dangerous
\item \code{5} not dangerous at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   2010 \tab A/B/- \tab partial \cr
   2016 \tab A/-/- \tab partial \cr
   2018 \tab A/-/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4756/vshow}{GSS Data Explorer page for \code{nukegen}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{carsfam}},
\code{\link{carsgen}},
\code{\link{carsten}},
\code{\link{grntest3}},
\code{\link{grntest4}},
\code{\link{grntest5}},
\code{\link{grntest6}},
\code{\link{grntest7}},
\code{\link{nukefam}}

Other Environment: 
\code{\link{carsfam}},
\code{\link{carsgen}},
\code{\link{carsten}},
\code{\link{grntest3}},
\code{\link{grntest4}},
\code{\link{grntest5}},
\code{\link{grntest6}},
\code{\link{grntest7}},
\code{\link{nukefam}}
}
\concept{Environment}
\concept{ISSP Environment}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
