% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_122.R
\name{numbwfrd}
\alias{numbwfrd}
\title{Good friends who are white-black}
\source{
General Social Survey https://gss.norc.org
}
\description{
numbwfrd
}
\details{
Question How many of your good friends are White/Black?
}
\section{Values}{

\itemize{
\item \code{94} no numuber given
\item \code{95} many
\item \code{96} 96+
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1135/vshow}{GSS Data Explorer page for \code{numbwfrd}}.

Counts by year:\tabular{llllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 15 \tab 2 \tab 20 \tab 25 \tab 27 \tab 3 \tab 30 \tab 35 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1976 \tab 580 \tab 117 \tab 8 \tab 1 \tab 1 \tab 1 \tab 1 \tab 62 \tab 3 \tab 1 \tab 1 \tab 26 \tab 1 \tab 1 \tab 12 \tab 10 \tab 4 \tab 4 \tab 1 \tab 1 \tab 20 \tab 2832 \cr
   Total \tab 1976 \tab 580 \tab 117 \tab 8 \tab 1 \tab 1 \tab 1 \tab 1 \tab 62 \tab 3 \tab 1 \tab 1 \tab 26 \tab 1 \tab 1 \tab 12 \tab 10 \tab 4 \tab 4 \tab 1 \tab 1 \tab 20 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{bwfriend}},
\code{\link{friends}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{givecong}},
\code{\link{godsells}},
\code{\link{godsport}},
\code{\link{numfrend}}
}
\concept{Religion}
\keyword{variable}
