% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_122.R
\name{numbwfrd}
\alias{numbwfrd}
\title{Good friends who are white-black}
\source{
General Social Survey https://gss.norc.org
}
\description{
numbwfrd
}
\details{
Question How many of your good friends are White/Black?
}
\section{Values}{

\itemize{
\item \code{94} no numuber given
\item \code{95} many
\item \code{96} 96+
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1135/vshow}{GSS Data Explorer page for \code{numbwfrd}}.
}

\seealso{
Other Religion: 
\code{\link{bwfriend}},
\code{\link{friends}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{givecong}},
\code{\link{godsells}},
\code{\link{godsport}},
\code{\link{numfrend}}

Other Friends: 
\code{\link{bwfriend}},
\code{\link{friends}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{numfrend}}

Other Race: 
\code{\link{bwfriend}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}}
}
\concept{Friends}
\concept{Race}
\concept{Religion}
\keyword{variable}
