% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_349.R
\name{numchurh}
\alias{numchurh}
\title{How many r belongs to church affiliate groups}
\source{
General Social Survey https://gss.norc.org
}
\description{
numchurh
}
\details{
Question O. How many church-affiliated groups do you belong to?
}
\section{Values}{

\itemize{
\item \code{7} seven or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3141/vshow}{GSS Data Explorer page for \code{numchurh}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{churchtx}},
\code{\link{frstmet1}},
\code{\link{frstmet2}},
\code{\link{frstmet3}},
\code{\link{frstmet4}},
\code{\link{frstmet5}},
\code{\link{grpboth1}},
\code{\link{grpboth2}},
\code{\link{grpboth3}},
\code{\link{grpboth4}},
\code{\link{grpboth5}},
\code{\link{infrmgrp}},
\code{\link{mtchurh1}},
\code{\link{mtchurh2}},
\code{\link{mtchurh3}},
\code{\link{mtchurh4}},
\code{\link{mtchurh5}},
\code{\link{mtchurh6}},
\code{\link{mtfarm1}},
\code{\link{mtfarm2}},
\code{\link{mtfarm3}},
\code{\link{mtfrat1}},
\code{\link{mtfrat2}},
\code{\link{mtfrat3}},
\code{\link{mtfrat4}},
\code{\link{mtgreek1}},
\code{\link{mtgreek2}},
\code{\link{mtgreek3}},
\code{\link{mthobby1}},
\code{\link{mthobby2}},
\code{\link{mthobby3}},
\code{\link{mtinfrm1}},
\code{\link{mtinfrm2}},
\code{\link{mtinfrm3}},
\code{\link{mtinfrm4}},
\code{\link{mtinfrm5}},
\code{\link{mtlit1}},
\code{\link{mtlit2}},
\code{\link{mtlit3}},
\code{\link{mtlit4}},
\code{\link{mtnat1}},
\code{\link{mtnat2}},
\code{\link{mtnat3}},
\code{\link{mtother1}},
\code{\link{mtother2}},
\code{\link{mtother3}},
\code{\link{mtother4}},
\code{\link{mtother5}},
\code{\link{mtpolit1}},
\code{\link{mtpolit2}},
\code{\link{mtpolit3}},
\code{\link{mtprof1}},
\code{\link{mtprof2}},
\code{\link{mtprof3}},
\code{\link{mtprof4}},
\code{\link{mtprof5}},
\code{\link{mtprof6}},
\code{\link{mtschl1}},
\code{\link{mtschl2}},
\code{\link{mtschl3}},
\code{\link{mtschl4}},
\code{\link{mtserv1}},
\code{\link{mtserv2}},
\code{\link{mtserv3}},
\code{\link{mtserv4}},
\code{\link{mtserv5}},
\code{\link{mtserv6}},
\code{\link{mtserv7}},
\code{\link{mtserv8}},
\code{\link{mtsport1}},
\code{\link{mtsport2}},
\code{\link{mtsport3}},
\code{\link{mtsport4}},
\code{\link{mtsport5}},
\code{\link{mtsport6}},
\code{\link{mtunion1}},
\code{\link{mtunion2}},
\code{\link{mtunion3}},
\code{\link{mtunion4}},
\code{\link{mtvet1}},
\code{\link{mtvet2}},
\code{\link{mtvet3}},
\code{\link{mtvet4}},
\code{\link{mtvet5}},
\code{\link{mtyouth1}},
\code{\link{mtyouth2}},
\code{\link{mtyouth3}},
\code{\link{mtyouth4}},
\code{\link{mtyouth5}},
\code{\link{numfarm}},
\code{\link{numfrat}},
\code{\link{numgreek}},
\code{\link{numhobby}},
\code{\link{numinfrm}},
\code{\link{numlit}},
\code{\link{numnat}},
\code{\link{numother}},
\code{\link{numpolit}},
\code{\link{numprof}},
\code{\link{numschl}},
\code{\link{numserv}},
\code{\link{numsport}},
\code{\link{numunion}},
\code{\link{numvet}},
\code{\link{numyouth}},
\code{\link{relneg}},
\code{\link{typchurh}},
\code{\link{typfarm}},
\code{\link{typfrat}},
\code{\link{typgreek}},
\code{\link{typhobby}},
\code{\link{typinfrm}},
\code{\link{typlit}},
\code{\link{typnat}},
\code{\link{typother}},
\code{\link{typpolit}},
\code{\link{typprof}},
\code{\link{typschl}},
\code{\link{typserv}},
\code{\link{typsport}},
\code{\link{typunion}},
\code{\link{typvet}},
\code{\link{typyouth}},
\code{\link{yrchurh1}},
\code{\link{yrchurh2}},
\code{\link{yrchurh3}},
\code{\link{yrchurh4}},
\code{\link{yrchurh5}},
\code{\link{yrchurh6}},
\code{\link{yrfarm1}},
\code{\link{yrfarm2}},
\code{\link{yrfarm3}},
\code{\link{yrfrat1}},
\code{\link{yrfrat2}},
\code{\link{yrfrat3}},
\code{\link{yrfrat4}},
\code{\link{yrgreek1}},
\code{\link{yrgreek2}},
\code{\link{yrgreek3}},
\code{\link{yrhobby1}},
\code{\link{yrhobby2}},
\code{\link{yrhobby3}},
\code{\link{yrinfrm1}},
\code{\link{yrinfrm2}},
\code{\link{yrinfrm3}},
\code{\link{yrinfrm4}},
\code{\link{yrinfrm5}},
\code{\link{yrlit1}},
\code{\link{yrlit2}},
\code{\link{yrlit3}},
\code{\link{yrlit4}},
\code{\link{yrnat1}},
\code{\link{yrnat2}},
\code{\link{yrnat3}},
\code{\link{yrother1}},
\code{\link{yrother2}},
\code{\link{yrother3}},
\code{\link{yrother4}},
\code{\link{yrother5}},
\code{\link{yrpolit1}},
\code{\link{yrpolit2}},
\code{\link{yrpolit3}},
\code{\link{yrprof1}},
\code{\link{yrprof2}},
\code{\link{yrprof3}},
\code{\link{yrprof4}},
\code{\link{yrprof5}},
\code{\link{yrprof6}},
\code{\link{yrschl1}},
\code{\link{yrschl2}},
\code{\link{yrschl3}},
\code{\link{yrschl4}},
\code{\link{yrserv1}},
\code{\link{yrserv2}},
\code{\link{yrserv3}},
\code{\link{yrserv4}},
\code{\link{yrserv5}},
\code{\link{yrserv6}},
\code{\link{yrserv7}},
\code{\link{yrserv8}},
\code{\link{yrskwn1}},
\code{\link{yrskwn2}},
\code{\link{yrskwn3}},
\code{\link{yrskwn4}},
\code{\link{yrskwn5}},
\code{\link{yrsport1}},
\code{\link{yrsport2}},
\code{\link{yrsport3}},
\code{\link{yrsport4}},
\code{\link{yrsport5}},
\code{\link{yrsport6}},
\code{\link{yrunion1}},
\code{\link{yrunion2}},
\code{\link{yrunion3}},
\code{\link{yrunion4}},
\code{\link{yrvet1}},
\code{\link{yrvet2}},
\code{\link{yrvet3}},
\code{\link{yrvet4}},
\code{\link{yrvet5}},
\code{\link{yryouth1}},
\code{\link{yryouth2}},
\code{\link{yryouth3}},
\code{\link{yryouth4}},
\code{\link{yryouth5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
