% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_279.R
\name{numcntct}
\alias{numcntct}
\title{How many people in contact least once a year}
\source{
General Social Survey https://gss.norc.org
}
\description{
numcntct
}
\details{
Question Not counting people at work or family at home, about how many other friends or relatives do you keep in contact with at least once a year?
}
\section{Values}{

\itemize{
\item \code{995} unspecified, 1 or more
\item \code{996} too many to list
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/- \tab partial \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2543/vshow}{GSS Data Explorer page for \code{numcntct}}.
}

\seealso{
Other Information Society: 
\code{\link{artgn}},
\code{\link{byemail}},
\code{\link{byphone}},
\code{\link{inperson}},
\code{\link{letters}},
\code{\link{meetings}},
\code{\link{perfrmgn}},
\code{\link{polcangn}},
\code{\link{polinfgn}}

Other Family: 
\code{\link{byemail}},
\code{\link{byphone}},
\code{\link{inperson}},
\code{\link{letters}},
\code{\link{meetings}}
}
\concept{Family}
\concept{Information Society}
\keyword{variable}
