% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_579.R
\name{numdays}
\alias{numdays}
\title{Number of days attended religious services}
\source{
General Social Survey https://gss.norc.org
}
\description{
numdays
}
\details{
Question On what day or days do you attend religious services during the last seven days? (PROBE, ASK UNTIL "NO" IS GIVEN: "Did you attend religious services on any other days during the last seven days?")
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/C/- \tab partial \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5310/vshow}{GSS Data Explorer page for \code{numdays}}.
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{attrelig}},
\code{\link{eatout}},
\code{\link{godoc}},
\code{\link{monday}},
\code{\link{seefilm}},
\code{\link{sunday}}

Other Religion: 
\code{\link{attrelig}},
\code{\link{monday}},
\code{\link{sunday}}
}
\concept{Religion}
\concept{Religious Attendance and Identity}
\keyword{variable}
