% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_240.R
\name{numemp}
\alias{numemp}
\title{For how many employers ever worked full time}
\source{
General Social Survey https://gss.norc.org
}
\description{
numemp
}
\details{
Question For how many employers have you ever worked full time?  READ CATEGORIES.
}
\section{Values}{

\itemize{
\item \code{1} one
\item \code{2} two
\item \code{3} three to five
\item \code{4} six to ten
\item \code{5} more than ten
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2159/vshow}{GSS Data Explorer page for \code{numemp}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab one \tab two \tab three to five \tab six to ten \tab more than ten \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1455 \tab 182 \tab 211 \tab 606 \tab 223 \tab 104 \tab 2 \tab 49 \tab 2832 \cr
   Total \tab 1455 \tab 182 \tab 211 \tab 606 \tab 223 \tab 104 \tab 2 \tab 49 \tab 2832 \cr
}
}

\seealso{
Other Jobs: 
\code{\link{doingdys}},
\code{\link{doingfin}},
\code{\link{doingtrn}},
\code{\link{formldys}},
\code{\link{formlfin}},
\code{\link{formlhrs}},
\code{\link{howpaid}},
\code{\link{wojob}},
\code{\link{wojobyrs}}
}
\concept{Jobs}
\keyword{variable}
