% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_502.R
\name{numemply}
\alias{numemply}
\title{Number of persons r employs}
\source{
General Social Survey https://gss.norc.org
}
\description{
numemply
}
\details{
Question How many people do you employ (include part-time and full-time employees?)
}
\section{Values}{

\itemize{
\item \code{1} 1-9
\item \code{2} 10-49
\item \code{3} 50-99
\item \code{4} 100-499
\item \code{5} 500-999
\item \code{6} 1,000-1,999
\item \code{7} 2,000+
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4570/vshow}{GSS Data Explorer page for \code{numemply}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab 1-9 \tab 10-49 \tab 50-99 \tab 100-499 \tab no answer \tab Total \cr
   1989 \tab 1477 \tab 30 \tab 17 \tab 2 \tab 1 \tab 10 \tab 1537 \cr
   Total \tab 1477 \tab 30 \tab 17 \tab 2 \tab 1 \tab 10 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{employer}},
\code{\link{localnum}},
\code{\link{moonlite}},
\code{\link{numsites}},
\code{\link{othhrs}},
\code{\link{supervis}},
\code{\link{supnum}},
\code{\link{suprvsjb}},
\code{\link{totalnum}}
}
\concept{Work Orientation}
\keyword{variable}
