% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_502.R
\name{numemply}
\alias{numemply}
\title{Number of persons r employs}
\source{
General Social Survey https://gss.norc.org
}
\description{
numemply
}
\details{
Question How many people do you employ (include part-time and full-time employees?)
}
\section{Values}{

\itemize{
\item \verb{[1]} 1-9
\item \verb{[2]} 10-49
\item \verb{[3]} 50-99
\item \verb{[4]} 100-499
\item \verb{[5]} 500-999
\item \verb{[6]} 1,000-1,999
\item \verb{[7]} 2,000+
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4570/vshow}{GSS Data Explorer page for \code{numemply}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab 1-9 \tab 10-49 \tab 50-99 \tab 100-499 \tab no answer \tab Total \cr
   1989 \tab 1477 \tab 30 \tab 17 \tab 2 \tab 1 \tab 10 \tab 1537 \cr
   Total \tab 1477 \tab 30 \tab 17 \tab 2 \tab 1 \tab 10 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{employer}},
\code{\link{localnum}},
\code{\link{moonlite}},
\code{\link{numsites}},
\code{\link{othhrs}},
\code{\link{supervis}},
\code{\link{supnum}},
\code{\link{suprvsjb}},
\code{\link{totalnum}}
}
\concept{Work Orientation}
\keyword{variable}
