% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_418.R
\name{numemps5}
\alias{numemps5}
\title{How many people r employs 5 yrs after}
\source{
General Social Survey https://gss.norc.org
}
\description{
numemps5
}
\details{
Question How many people do you anticipate employing five years from now?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3823/vshow}{GSS Data Explorer page for \code{numemps5}}.

Counts by year:\tabular{lllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 100 \tab 12 \tab 15 \tab 2 \tab 20 \tab 22 \tab 3 \tab 30 \tab 4 \tab 40 \tab 5 \tab 50 \tab 55 \tab 6 \tab 7 \tab 70 \tab 8 \tab don't know \tab Total \cr
   2008 \tab 1882 \tab 19 \tab 12 \tab 3 \tab 2 \tab 3 \tab 23 \tab 8 \tab 1 \tab 11 \tab 1 \tab 7 \tab 2 \tab 21 \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab 4 \tab 14 \tab 2023 \cr
   Total \tab 1882 \tab 19 \tab 12 \tab 3 \tab 2 \tab 3 \tab 23 \tab 8 \tab 1 \tab 11 \tab 1 \tab 7 \tab 2 \tab 21 \tab 2 \tab 1 \tab 5 \tab 1 \tab 1 \tab 4 \tab 14 \tab 2023 \cr
}
}

\seealso{
Other Self Employment: 
\code{\link{bizgross}},
\code{\link{bizshare}},
\code{\link{owninc}},
\code{\link{paidemps}},
\code{\link{whybiz}}
}
\concept{Self Employment}
\keyword{variable}
