% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_686.R
\name{numemps_0422}
\alias{numemps_0422}
\title{If self employed, number of other employees}
\source{
General Social Survey https://gss.norc.org
}
\description{
numemps_0422
}
\details{
Question Are/were you self-employed or do/did you work for someone else?
(If "Self-employed") Do/did you have any employees?
(If "Yes") How many employees do/did you have?

As of 2024, NUMEMPS has been recoded to ranges of employees. This has been retroactively applied to all previous years of NUMEMPS. The original NUMEMPS containing data from 2004 to 2022 has been renamed NUMEMPS_0422.
}
\section{Values}{

\itemize{
\item \code{97} 97 or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8152/vshow}{GSS Data Explorer page for \code{numemps_0422}}.
}

\concept{Business}
\concept{Employment}
\concept{Workplace Conflict}
\keyword{variable}
