% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_122.R
\name{numfrend}
\alias{numfrend}
\title{About how many good friends do you have}
\source{
General Social Survey https://gss.norc.org
}
\description{
numfrend
}
\details{
Question B. About how many good friends do you have?
}
\section{Values}{

\itemize{
\item \code{94} no number given
\item \code{95} many
\item \code{96} 96+
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1134/vshow}{GSS Data Explorer page for \code{numfrend}}.
}

\seealso{
Other Religion: 
\code{\link{bwfriend}},
\code{\link{friends}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{givecong}},
\code{\link{godsells}},
\code{\link{godsport}},
\code{\link{numbwfrd}}
}
\concept{Religion}
\keyword{variable}
