% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_93.R
\name{numgiven}
\alias{numgiven}
\title{Number of persons mentioned}
\source{
General Social Survey https://gss.norc.org
}
\description{
numgiven
}
\details{
Question From time to time, most people discuss important matters with other people. Looking back over the last six months - who are the people with whom you discussed matters important to you? Just tell me their first names or initials. If LESS THAN 5 NAMES MENTIONED, PROBE, Anyone else?  ONLY RECORD FIRST 5 NAMES.
LIST ALL NAMES IN ORDER ACROSS THE TOP OF THE MATRIX (SEE 2 PAGES AHEAD). THEN WRITE NAMES 2-5 DOWN THE SIDE OF THE MATRIX.
A. INTERVIEWER CHECK: HOW MANY NAMES WERE MENTIONED?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/848/vshow}{GSS Data Explorer page for \code{numgiven}}.
}

\seealso{
Other Social Networks: 
\code{\link{eqclose}}
}
\concept{Social Networks}
\keyword{variable}
