% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_349.R
\name{numinfrm}
\alias{numinfrm}
\title{How many r belongs to informal groups}
\source{
General Social Survey https://gss.norc.org
}
\description{
numinfrm
}
\details{
Question Q. How many of informal groups do you belong to?
}
\section{Values}{

\itemize{
\item \code{7} seven or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3143/vshow}{GSS Data Explorer page for \code{numinfrm}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{numchurh}},
\code{\link{numfarm}},
\code{\link{numgreek}},
\code{\link{numhobby}},
\code{\link{numlit}},
\code{\link{numnat}},
\code{\link{numother}},
\code{\link{numprof}},
\code{\link{numschl}}

Other Voluntary Associations: 
\code{\link{numchurh}},
\code{\link{numfarm}},
\code{\link{numgreek}},
\code{\link{numhobby}},
\code{\link{numlit}},
\code{\link{numnat}},
\code{\link{numother}},
\code{\link{numprof}},
\code{\link{numschl}}
}
\concept{Voluntary Assocations/Social Networks}
\concept{Voluntary Associations}
\keyword{variable}
