% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_480.R
\name{numkids}
\alias{numkids}
\title{What is ideal number of kids for family}
\source{
General Social Survey https://gss.norc.org
}
\description{
numkids
}
\details{
Question All in all, what do you think is the ideal number of children for a family to have?  Please just write a number in the box below.
}
\section{Values}{

\itemize{
\item \code{7} seven or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   2012 \tab A/B/- \tab partial \cr
   2022 \tab A/-/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4374/vshow}{GSS Data Explorer page for \code{numkids}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{fourkids}},
\code{\link{kidtrble}},
\code{\link{mardiv}},
\code{\link{nokids}},
\code{\link{onekid}},
\code{\link{threkids}},
\code{\link{twokids}}

Other Children: 
\code{\link{fourkids}},
\code{\link{kidtrble}},
\code{\link{nokids}},
\code{\link{onekid}},
\code{\link{threkids}},
\code{\link{twokids}}
}
\concept{Children}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
