% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_393.R
\name{numknown}
\alias{numknown}
\title{Variable that drives skips in module}
\source{
General Social Survey https://gss.norc.org
}
\description{
numknown
}
\details{
Question Cases in Sample B (Ballots 4-6) should receive this variable. This variable is used to drive skips in a module. This  variable should have a value of 1-5 using the following proportions: 1=50pct, 2,3,4,5 = 12.5pct each
}
\section{Values}{

\itemize{
\item \code{1} subset1
\item \code{2} subset2
\item \code{3} subset3
\item \code{4} subset4
\item \code{5} subset5
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3571/vshow}{GSS Data Explorer page for \code{numknown}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab subset1 \tab subset2 \tab subset3 \tab subset4 \tab subset5 \tab Total \cr
   2006 \tab 3041 \tab 741 \tab 178 \tab 188 \tab 190 \tab 172 \tab 4510 \cr
   Total \tab 3041 \tab 741 \tab 178 \tab 188 \tab 190 \tab 172 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{acqbrnda}},
\code{\link{acqkaren}},
\code{\link{acqkeith}},
\code{\link{acqkevin}},
\code{\link{acqmark}},
\code{\link{acqrachl}},
\code{\link{acqshawn}}
}
\concept{Number Known}
\keyword{variable}
