% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_466.R
\name{numlangs}
\alias{numlangs}
\title{How many languages does r speak well enough to hold a conversation in?}
\source{
General Social Survey https://gss.norc.org
}
\description{
numlangs
}
\details{
Question How many languages do you speak well enough to hold a conversation in, including the language(s) you speak at home?
}
\section{Values}{

\itemize{
\item \code{1} one language
\item \code{2} two languages
\item \code{3} three languages
\item \code{4} four or more languages
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7277/vshow}{GSS Data Explorer page for \code{numlangs}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab one language \tab two languages \tab three languages \tab four or more languages \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1175 \tab 920 \tab 213 \tab 29 \tab 7 \tab 3 \tab 1 \tab 2348 \cr
   Total \tab 1175 \tab 920 \tab 213 \tab 29 \tab 7 \tab 3 \tab 1 \tab 2348 \cr
}
}

\keyword{variable}
