% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_553.R
\name{nummen}
\alias{nummen}
\title{Number of male sex partner's since 18}
\source{
General Social Survey https://gss.norc.org
}
\description{
nummen
}
\details{
Question Now thinking about the time since your 18th birthday (including the past 12 months) how many male partners have you had sex with?
}
\section{Values}{

\itemize{
\item \code{989} 989 or more
\item \code{990} dash or slash
\item \code{991} 1 or more, number unknown
\item \code{992} 'x'
\item \code{993} garbled text
\item \code{994} 'several'
\item \code{995} 'many,' 'lots'
\item \code{996} 'n.a.'
\item \code{997} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab B/C/- \tab partial \cr
   2018 \tab B/C/- \tab partial \cr
   2021 \tab B/C/- \tab partial \cr
   2022 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5061/vshow}{GSS Data Explorer page for \code{nummen}}.
}

\seealso{
Other High Risk Behaviors: 
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{numwomen}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{sexsex5}}
}
\concept{High Risk Behaviors}
\keyword{variable}
