% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_391.R
\name{numorg}
\alias{numorg}
\title{Number of people working in organization at all locations}
\source{
General Social Survey https://gss.norc.org
}
\description{
numorg
}
\details{
Question About how many people work in your organization at all locations? Count part-time as well as full-time employees.
}
\section{Values}{

\itemize{
\item \code{9981} 1-9 in range
\item \code{9982} 10-49 in range
\item \code{9983} 50-99 in range
\item \code{9984} 100-499 in range
\item \code{9985} 500-999 in range
\item \code{9986} 1,000-1,999 in range
\item \code{9987} 2,000-2,999 in range
\item \code{9988} 3,000-3,999 in range
\item \code{9989} 4,000-4,999 in range
\item \code{9990} 5,000-10,000 in range
\item \code{9991} 10,000-50,000 in range
\item \code{9992} 50,000-100,000 in range
\item \code{9993} above 100,000 in range
\item \code{9997} 10,000 or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3556/vshow}{GSS Data Explorer page for \code{numorg}}.
}

\seealso{
Other Shared Capitalism: 
\code{\link{buyvalue}},
\code{\link{compwage}},
\code{\link{defpensn}},
\code{\link{empinput}},
\code{\link{emptrain}},
\code{\link{esop}},
\code{\link{extr2005}},
\code{\link{ops2005}},
\code{\link{slfmangd}},
\code{\link{talkteam}},
\code{\link{trdestck}},
\code{\link{wealth}}
}
\concept{Shared Capitalism}
\keyword{variable}
