% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_652.R
\name{numorg1}
\alias{numorg1}
\title{Estimate of total people in r's business or organization}
\source{
General Social Survey https://gss.norc.org
}
\description{
numorg1
}
\details{
Question About how many people work in your organization at all locations? Count part-time as well as full-time employees.
}
\section{Values}{

\itemize{
\item \verb{[1]} 1-9 in range
\item \verb{[2]} 10-49 in range
\item \verb{[3]} 50-99 in range
\item \verb{[4]} 100-499 in range
\item \verb{[5]} 500-999 in range
\item \verb{[6]} 1,000-1,999 in range
\item \verb{[7]} 2,000-2,999 in range
\item \verb{[8]} 3,000-3,999 in range
\item \verb{[9]} 4,000-4,999 in range
\item \verb{[10]} 5,000-10,000 in range
\item \verb{[11]} 10,000-50,000 in range
\item \verb{[12]} 50,000-100,000 in range
\item \verb{[13]} above 100,000 in range
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7852/vshow}{GSS Data Explorer page for \code{numorg1}}.

Counts by year:\tabular{lllllllllllllllllll}{
   year \tab iap \tab 1-9 in range \tab 10-49 in range \tab 50-99 in range \tab 100-499 in range \tab 500-999 in range \tab 1,000-1,999 in range \tab 2,000-2,999 in range \tab 3,000-3,999 in range \tab 4,000-4,999 in range \tab 5,000-10,000 in range \tab 10,000-50,000 in range \tab 50,000-100,000 in range \tab above 100,000 in range \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2569 \tab 154 \tab 146 \tab 67 \tab 127 \tab 57 \tab 39 \tab 28 \tab 15 \tab 15 \tab 38 \tab 49 \tab 30 \tab 63 \tab 127 \tab 10 \tab 10 \tab 3544 \cr
   Total \tab 2569 \tab 154 \tab 146 \tab 67 \tab 127 \tab 57 \tab 39 \tab 28 \tab 15 \tab 15 \tab 38 \tab 49 \tab 30 \tab 63 \tab 127 \tab 10 \tab 10 \tab 3544 \cr
}
}

\keyword{variable}
