% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_349.R
\name{numother}
\alias{numother}
\title{How many r belongs to other groups}
\source{
General Social Survey https://gss.norc.org
}
\description{
numother
}
\details{
Question P. How many of any other groups do you belong to?
}
\section{Values}{

\itemize{
\item \code{7} seven or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3142/vshow}{GSS Data Explorer page for \code{numother}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab 5 \tab no answer \tab Total \cr
   2004 \tab 2706 \tab 88 \tab 6 \tab 1 \tab 2 \tab 9 \tab 2812 \cr
   Total \tab 2706 \tab 88 \tab 6 \tab 1 \tab 2 \tab 9 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{numchurh}},
\code{\link{numfarm}},
\code{\link{numgreek}},
\code{\link{numhobby}},
\code{\link{numinfrm}},
\code{\link{numlit}},
\code{\link{numnat}},
\code{\link{numprof}},
\code{\link{numschl}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
