% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_417.R
\name{numown}
\alias{numown}
\title{Number of businesses you own and manage}
\source{
General Social Survey https://gss.norc.org
}
\description{
numown
}
\details{
Question In how many privately-held businesses, professional practices, limited partnerships, or any other types of partnership do you own or share ownership in and have an active management role?
}
\section{Values}{

\itemize{
\item \verb{[7]} 7 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3816/vshow}{GSS Data Explorer page for \code{numown}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab don't know \tab Total \cr
   2008 \tab 1774 \tab 205 \tab 24 \tab 10 \tab 5 \tab 1 \tab 3 \tab 1 \tab 2023 \cr
   Total \tab 1774 \tab 205 \tab 24 \tab 10 \tab 5 \tab 1 \tab 3 \tab 1 \tab 2023 \cr
}
}

\seealso{
Other Self Employment: 
\code{\link{ownbiz}},
\code{\link{startbiz}},
\code{\link{yearbiz}}
}
\concept{Self Employment}
\keyword{variable}
