% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_317.R
\name{numpets}
\alias{numpets}
\title{How many pets r's family has}
\source{
General Social Survey https://gss.norc.org
}
\description{
numpets
}
\details{
Question How many pets does your family have? By family, we mean members of your family with whom you currently live.
}
\section{Values}{

\itemize{
\item \verb{[0]} no pets
\item \verb{[20]} 20 or more pets
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7194/vshow}{GSS Data Explorer page for \code{numpets}}.

Counts by year:\tabular{lllllllllllllllllllllllll}{
   year \tab iap \tab no pets \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 or more pets \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1175 \tab 474 \tab 288 \tab 4 \tab 7 \tab 5 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 159 \tab 12 \tab 82 \tab 45 \tab 29 \tab 14 \tab 8 \tab 5 \tab 5 \tab 1 \tab 25 \tab 2348 \cr
   Total \tab 1175 \tab 474 \tab 288 \tab 4 \tab 7 \tab 5 \tab 3 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 159 \tab 12 \tab 82 \tab 45 \tab 29 \tab 14 \tab 8 \tab 5 \tab 5 \tab 1 \tab 25 \tab 2348 \cr
}
}

\keyword{variable}
