% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_288.R
\name{numprobs}
\alias{numprobs}
\title{How many friends close to discuss problems}
\source{
General Social Survey https://gss.norc.org
}
\description{
numprobs
}
\details{
Question Of these (NUMCNTCT) friends and relatives, about how many would you say you feel really close to, that is close enough to discuss personal or important problems with?
}
\section{Values}{

\itemize{
\item \code{96} 96 or higher
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2606/vshow}{GSS Data Explorer page for \code{numprobs}}.
}

\seealso{
Other Information Society: 
\code{\link{byemprob}},
\code{\link{intaccss}},
\code{\link{jobinfo}},
\code{\link{jobpaper}}

Other Family: 
\code{\link{byemprob}}
}
\concept{Family}
\concept{Information Society}
\keyword{variable}
