% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_139.R
\name{numpromo}
\alias{numpromo}
\title{#Promotions have you received}
\source{
General Social Survey https://gss.norc.org
}
\description{
numpromo
}
\details{
Question Have you received any promotions while working for your present employer (IF SELF-EMPLOYED:  current organization)?
A. How many promotions have you received?
}
\section{Values}{

\itemize{
\item \verb{[7]} 7 or more
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1301/vshow}{GSS Data Explorer page for \code{numpromo}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 or more \tab don't know \tab no answer \tab Total \cr
   1991 \tab 1148 \tab 132 \tab 61 \tab 41 \tab 26 \tab 5 \tab 3 \tab 10 \tab 1 \tab 90 \tab 1517 \cr
   Total \tab 1148 \tab 132 \tab 61 \tab 41 \tab 26 \tab 5 \tab 3 \tab 10 \tab 1 \tab 90 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{advances}},
\code{\link{futpromo}},
\code{\link{imppromo}},
\code{\link{jobcntrl}},
\code{\link{lottosay}},
\code{\link{promoted}},
\code{\link{racpromo}},
\code{\link{sexpromo}},
\code{\link{wrkindep}}
}
\concept{Work Organizations}
\keyword{variable}
