% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_502.R
\name{numsites}
\alias{numsites}
\title{Does r's firm have more than one location}
\source{
General Social Survey https://gss.norc.org
}
\description{
numsites
}
\details{
Question Does your company or the institution for which you work have more than one physical location (that is, other offices, branches, plants, stores, or subsidiaries)? (If you work for the government, does your agency or division have more than one physical location?)?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4566/vshow}{GSS Data Explorer page for \code{numsites}}.

Counts by year:\tabular{lllllll}{
   year \tab not available in this year \tab yes \tab no \tab no answer \tab iap \tab Total \cr
   1989 \tab - \tab 518 \tab 267 \tab 752 \tab - \tab 1537 \cr
   Total \tab 0 \tab 518 \tab 267 \tab 752 \tab 0 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{employer}},
\code{\link{localnum}},
\code{\link{moonlite}},
\code{\link{numemply}},
\code{\link{othhrs}},
\code{\link{supervis}},
\code{\link{supnum}},
\code{\link{suprvsjb}},
\code{\link{totalnum}}
}
\concept{Work Orientation}
\keyword{variable}
