% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_348.R
\name{numunion}
\alias{numunion}
\title{How many r belongs to labor unions}
\source{
General Social Survey https://gss.norc.org
}
\description{
numunion
}
\details{
Question E. How many labor unions do you belong to?
}
\section{Values}{

\itemize{
\item \code{7} seven or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3131/vshow}{GSS Data Explorer page for \code{numunion}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{numfrat}},
\code{\link{numpolit}},
\code{\link{numserv}},
\code{\link{numsport}},
\code{\link{numvet}},
\code{\link{numyouth}},
\code{\link{typchurh}},
\code{\link{typinfrm}},
\code{\link{typother}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
