% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_348.R
\name{numvet}
\alias{numvet}
\title{How many r belongs to veterans groups}
\source{
General Social Survey https://gss.norc.org
}
\description{
numvet
}
\details{
Question C. How many veterans' groups do you belong to?
}
\section{Values}{

\itemize{
\item \code{7} seven or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3129/vshow}{GSS Data Explorer page for \code{numvet}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab no answer \tab Total \cr
   2004 \tab 2729 \tab 67 \tab 6 \tab 1 \tab 1 \tab 1 \tab 7 \tab 2812 \cr
   Total \tab 2729 \tab 67 \tab 6 \tab 1 \tab 1 \tab 1 \tab 7 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{numfrat}},
\code{\link{numpolit}},
\code{\link{numserv}},
\code{\link{numsport}},
\code{\link{numunion}},
\code{\link{numyouth}},
\code{\link{typchurh}},
\code{\link{typinfrm}},
\code{\link{typother}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
