% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_434.R
\name{obeylaw}
\alias{obeylaw}
\title{Should people obey laws without exceptn?}
\source{
General Social Survey https://gss.norc.org
}
\description{
obeylaw
}
\details{
Question In general, would you say that people should obey the law without exception, or are there exceptional occasions on which people should follow their consciences even if it means breaking the law? (CIRCLE ONE ANSWER)
}
\section{Values}{

\itemize{
\item \code{1} obey the law without exception
\item \code{2} follow conscience on occassions
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3981/vshow}{GSS Data Explorer page for \code{obeylaw}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{protest1}},
\code{\link{protest2}},
\code{\link{protest3}},
\code{\link{protest4}},
\code{\link{protest5}},
\code{\link{protest6}},
\code{\link{pubdef}},
\code{\link{pubecon}},
\code{\link{ratetone}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
