% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_125.R
\name{obeytch}
\alias{obeytch}
\title{Important to obey church teachings}
\source{
General Social Survey https://gss.norc.org
}
\description{
obeytch
}
\details{
Question Please tell me whether you strongly agree, agree, disagree, or strongly disagree with the following statements:
C. It’s important to obey church teachings even if I don’t understand them.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1160/vshow}{GSS Data Explorer page for \code{obeytch}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2135 \tab 37 \tab 222 \tab 300 \tab 59 \tab 40 \tab 24 \tab 2817 \cr
   Total \tab 2135 \tab 37 \tab 222 \tab 300 \tab 59 \tab 40 \tab 24 \tab 2817 \cr
}
}

\seealso{
Other Religion: 
\code{\link{ferespct}},
\code{\link{preach}},
\code{\link{sympcoun}},
\code{\link{worship}},
\code{\link{wrkyoung}}
}
\concept{Religion}
\keyword{variable}
