% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_518.R
\name{obeythnk}
\alias{obeythnk}
\title{Should chldrn be obedient or think for selves}
\source{
General Social Survey https://gss.norc.org
}
\description{
obeythnk
}
\details{
Question Which of these would you say is more important in preparing children for life…
}
\section{Values}{

\itemize{
\item \verb{[1]} to be obedient
\item \verb{[2]} to think for themselves
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4717/vshow}{GSS Data Explorer page for \code{obeythnk}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab to be obedient \tab to think for themselves \tab don't know \tab no answer \tab Total \cr
   1993 \tab 49 \tab 447 \tab 997 \tab 93 \tab 20 \tab 1606 \cr
   1994 \tab 159 \tab 815 \tab 1788 \tab 142 \tab 88 \tab 2992 \cr
   Total \tab 208 \tab 1262 \tab 2785 \tab 235 \tab 108 \tab 4598 \cr
}
}

\concept{ISSP Environment}
\keyword{variable}
