% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_518.R
\name{obeythnk}
\alias{obeythnk}
\title{Should chldrn be obedient or think for selves}
\source{
General Social Survey https://gss.norc.org
}
\description{
obeythnk
}
\details{
Question Which of these would you say is more important in preparing children for life…
}
\section{Values}{

\itemize{
\item \code{1} to be obedient
\item \code{2} to think for themselves
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4717/vshow}{GSS Data Explorer page for \code{obeythnk}}.
}

\concept{ISSP Environment}
\keyword{variable}
