% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_93.R
\name{obmewar}
\alias{obmewar}
\title{Obligation--wartime mil service-men}
\source{
General Social Survey https://gss.norc.org
}
\description{
obmewar
}
\details{
Question And one last question. We all know that American citizens have certain rights. For example, they have the right to free public education and to police protection, the right to attend religious services of their choice, and the right to elec public officials. I'd like to ask now about certain obligations that some people feel American citizens owe their country. I just want your own opinion on these--whether you feel it is a very important obligation, a somewhat important obligation, or not an obligation that a citizen owes to the country. (READ EACH STATEMENT AND CIRCLE ONE CODE for each).
(REPEAT ANSWER CATEGORIES AS NECESSARY)
H. For young men, serving in the military when the country is at war?
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} somewhat important
\item \verb{[3]} not obligation
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/841/vshow}{GSS Data Explorer page for \code{obmewar}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very important \tab somewhat important \tab not obligation \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 1208 \tab 198 \tab 34 \tab 12 \tab 21 \tab 1473 \cr
   Total \tab 0 \tab 1208 \tab 198 \tab 34 \tab 12 \tab 21 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{helpblk}},
\code{\link{helpnot}},
\code{\link{helppoor}},
\code{\link{helpsick}},
\code{\link{obfepax}},
\code{\link{obfewar}},
\code{\link{obmepax}}
}
\concept{Core}
\keyword{variable}
