% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_175.R
\name{obrespct}
\alias{obrespct}
\title{Teach children respect for authority and obedience}
\source{
General Social Survey https://gss.norc.org
}
\description{
obrespct
}
\details{
Question Please respond to the following statements by saying whether you strongly agree, agree, disagree or strongly disagree with them:
Obedience and respect for authority are the most important virtues children should learn.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1640/vshow}{GSS Data Explorer page for \code{obrespct}}.
}

\seealso{
Other Multiculturalism: 
\code{\link{asngovt}},
\code{\link{blkgovt}},
\code{\link{econpast}},
\code{\link{hspgovt}},
\code{\link{whtgovt}},
\code{\link{wlthimm}},
\code{\link{wlthundc}},
\code{\link{workimm}},
\code{\link{workundc}}
}
\concept{Children}
\concept{Multiculturalism}
\keyword{variable}
