% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_147.R
\name{obtohelp}
\alias{obtohelp}
\title{Obligation to care for ill parent or spouse?}
\source{
General Social Survey https://gss.norc.org
}
\description{
obtohelp
}
\details{
Question B. Sometimes people are asked to make great sacrifices for others in their family by caring for an elderly parent or a spouse who is seriously ill or injured. When someone is asked to give such care, is it their obligation to do so, or should they give care only if they really want to?
}
\section{Values}{

\itemize{
\item \code{1} it is oblgation
\item \code{2} only if they want to
\item \code{3} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1381/vshow}{GSS Data Explorer page for \code{obtohelp}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab it is oblgation \tab only if they want to \tab other \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 890 \tab 626 \tab 57 \tab 24 \tab 9 \tab 1606 \cr
   Total \tab 0 \tab 890 \tab 626 \tab 57 \tab 24 \tab 9 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{bigband}},
\code{\link{blugrass}},
\code{\link{lfechnce}},
\code{\link{lfegenes}},
\code{\link{lfegod}},
\code{\link{lfehrdwk}},
\code{\link{lfesocty}},
\code{\link{opoutcme}},
\code{\link{united}}
}
\concept{Culture}
\keyword{variable}
