% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_01.R
\name{occ}
\alias{occ}
\title{R's census occupation code (1970)}
\source{
General Social Survey https://gss.norc.org
}
\description{
occ
}
\details{
Question A. What kind of work do you (did you normally) do?  That is, what (is/was) your job called?
OCCUPATION:  (See REMARKS)
B. What (do/did) you actually do in that job?  Tell me, what (are/were) some of your main duties?
C. What kind of place (do/did) you work for?
INDUSTRY:  (See REMARKS)
D. What (do/did) they (make/do)?
E. (Are/Were) you self employed or (do/did) you work for someone else?
F. (Are/Were) you employed by the federal, state or local government or by a private employer (including non-profit organizations)?
G. About how much time (does/did) it usually take you to travel to work - about how many minutes?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\seealso{
Other Core: 
\code{\link{evwork}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{id}()},
\code{\link{prestige}},
\code{\link{wrkgovt}},
\code{\link{wrkslf}},
\code{\link{wrkstat}},
\code{\link{year}()}
}
\concept{Core}
\keyword{variable}
