% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_474.R
\name{occmobil}
\alias{occmobil}
\title{Status of job higher than father's?}
\source{
General Social Survey https://gss.norc.org
}
\description{
occmobil
}
\details{
Question Please think of your present job (or your last one if you don't have one now). If you compare this job with the job your father had when you were 16, would you say that the level or status of your job is (or was) . . .
}
\section{Values}{

\itemize{
\item \verb{[1]} much higher than dad
\item \verb{[2]} higher
\item \verb{[3]} about equal
\item \verb{[4]} lower
\item \verb{[5]} much lower than dad
\item \verb{[6]} r never had job
\item \verb{[7]} no dad-never had job
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   Full \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4313/vshow}{GSS Data Explorer page for \code{occmobil}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab much higher than dad \tab higher \tab about equal \tab lower \tab much lower than dad \tab r never had job \tab no dad-never had job \tab no answer \tab Total \cr
   1987 \tab 255 \tab 311 \tab 404 \tab 335 \tab 208 \tab 109 \tab 43 \tab 84 \tab 70 \tab 1819 \cr
   2000 \tab 1545 \tab 215 \tab 302 \tab 282 \tab 183 \tab 100 \tab 12 \tab 48 \tab 130 \tab 2817 \cr
   Total \tab 1800 \tab 526 \tab 706 \tab 617 \tab 391 \tab 209 \tab 55 \tab 132 \tab 200 \tab 4636 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conage}},
\code{\link{firstjob}},
\code{\link{firstslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{pajob}},
\code{\link{paslf}},
\code{\link{rewrdeff}},
\code{\link{rewrdint}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
