% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_238.R
\name{occtrain}
\alias{occtrain}
\title{Gone thru specialized training}
\source{
General Social Survey https://gss.norc.org
}
\description{
occtrain
}
\details{
Question A. Have you gone through any specialized training in school to obtain a vocational certificate, diploma, degree, or license for this occupation?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2138/vshow}{GSS Data Explorer page for \code{occtrain}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1998 \tab 1455 \tab 504 \tab 739 \tab 134 \tab 2832 \cr
   Total \tab 1455 \tab 504 \tab 739 \tab 134 \tab 2832 \cr
}
}

\seealso{
Other Jobs: 
\code{\link{occyrs}},
\code{\link{trainsch}},
\code{\link{typeorg}}
}
\concept{Jobs}
\keyword{variable}
