% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_238.R
\name{occyrs}
\alias{occyrs}
\title{How many years in your occupation}
\source{
General Social Survey https://gss.norc.org
}
\description{
occyrs
}
\details{
Question How long have you been doing the kind of work you did last week (OCCUPATION DESCRIBED IN OCC )?  That is, at least how many years in your occupation?
Please include all previous jobs of the same kind.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2137/vshow}{GSS Data Explorer page for \code{occyrs}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 3 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 4 \tab 40 \tab 41 \tab 42 \tab 44 \tab 45 \tab 46 \tab 48 \tab 5 \tab 50 \tab 54 \tab 55 \tab 58 \tab 6 \tab 60 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1455 \tab 131 \tab 89 \tab 87 \tab 20 \tab 33 \tab 27 \tab 25 \tab 58 \tab 16 \tab 18 \tab 32 \tab 13 \tab 101 \tab 67 \tab 11 \tab 16 \tab 7 \tab 11 \tab 28 \tab 7 \tab 8 \tab 12 \tab 5 \tab 97 \tab 29 \tab 5 \tab 6 \tab 2 \tab 9 \tab 7 \tab 1 \tab 3 \tab 4 \tab 71 \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 99 \tab 1 \tab 1 \tab 1 \tab 1 \tab 53 \tab 1 \tab 41 \tab 51 \tab 33 \tab 2 \tab 26 \tab 2832 \cr
   Total \tab 1455 \tab 131 \tab 89 \tab 87 \tab 20 \tab 33 \tab 27 \tab 25 \tab 58 \tab 16 \tab 18 \tab 32 \tab 13 \tab 101 \tab 67 \tab 11 \tab 16 \tab 7 \tab 11 \tab 28 \tab 7 \tab 8 \tab 12 \tab 5 \tab 97 \tab 29 \tab 5 \tab 6 \tab 2 \tab 9 \tab 7 \tab 1 \tab 3 \tab 4 \tab 71 \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 99 \tab 1 \tab 1 \tab 1 \tab 1 \tab 53 \tab 1 \tab 41 \tab 51 \tab 33 \tab 2 \tab 26 \tab 2832 \cr
}
}

\seealso{
Other Jobs: 
\code{\link{occtrain}},
\code{\link{trainsch}},
\code{\link{typeorg}}
}
\concept{Jobs}
\keyword{variable}
