% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_238.R
\name{occyrs}
\alias{occyrs}
\title{How many years in your occupation}
\source{
General Social Survey https://gss.norc.org
}
\description{
occyrs
}
\details{
Question How long have you been doing the kind of work you did last week (OCCUPATION DESCRIBED IN OCC )?  That is, at least how many years in your occupation?
Please include all previous jobs of the same kind.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2137/vshow}{GSS Data Explorer page for \code{occyrs}}.
}

\seealso{
Other Jobs: 
\code{\link{occtrain}},
\code{\link{trainsch}},
\code{\link{typeorg}}
}
\concept{Jobs}
\keyword{variable}
