% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_368.R
\name{odds1}
\alias{odds1}
\title{Test of knowledge about probablity1}
\source{
General Social Survey https://gss.norc.org
}
\description{
odds1
}
\details{
Question Now, think about this situation. A doctor tells a couple that their genetic makeup means that they’ve got one in four chances of having a child with an inherited illness.
A. Does this mean that if their first child has the illness, the next three will not have the illness?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3332/vshow}{GSS Data Explorer page for \code{odds1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 169 \tab 1596 \tab 96 \tab 3 \tab 4510 \cr
   2008 \tab 518 \tab 156 \tab 1228 \tab 108 \tab 13 \tab 2023 \cr
   2010 \tab 1328 \tab 75 \tab 585 \tab 48 \tab 8 \tab 2044 \cr
   2012 \tab 972 \tab 90 \tab 819 \tab 84 \tab 9 \tab 1974 \cr
   2014 \tab 1299 \tab 126 \tab 1021 \tab 88 \tab 4 \tab 2538 \cr
   2016 \tab 1477 \tab 161 \tab 1155 \tab 63 \tab 11 \tab 2867 \cr
   2018 \tab 1173 \tab 132 \tab 975 \tab 63 \tab 5 \tab 2348 \cr
   Total \tab 9413 \tab 909 \tab 7379 \tab 550 \tab 53 \tab 18304 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{astrolgy}},
\code{\link{astrosci}},
\code{\link{balneg}},
\code{\link{balpos}},
\code{\link{expdesgn}},
\code{\link{exptext}},
\code{\link{scibnfts}},
\code{\link{scistudy}},
\code{\link{scitext}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
