% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_369.R
\name{odds2}
\alias{odds2}
\title{Test of knowledge about probablity2}
\source{
General Social Survey https://gss.norc.org
}
\description{
odds2
}
\details{
Question Now, think about this situation. A doctor tells a couple that their genetic makeup means that they've got one in four chances of having a child with an inherited illness.
B. Does this mean that each of the couple's children will have the same risk of suffering from the illness?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3333/vshow}{GSS Data Explorer page for \code{odds2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 1408 \tab 356 \tab 96 \tab 4 \tab 4510 \cr
   2008 \tab 518 \tab 1079 \tab 311 \tab 101 \tab 14 \tab 2023 \cr
   2010 \tab 1328 \tab 538 \tab 130 \tab 39 \tab 9 \tab 2044 \cr
   2012 \tab 972 \tab 700 \tab 213 \tab 81 \tab 8 \tab 1974 \cr
   2014 \tab 1299 \tab 885 \tab 274 \tab 78 \tab 2 \tab 2538 \cr
   2016 \tab 1477 \tab 1010 \tab 317 \tab 50 \tab 13 \tab 2867 \cr
   2018 \tab 1173 \tab 854 \tab 261 \tab 54 \tab 6 \tab 2348 \cr
   Total \tab 9413 \tab 6474 \tab 1862 \tab 499 \tab 56 \tab 18304 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{bigbang}},
\code{\link{bigbang1}},
\code{\link{bigbang2}},
\code{\link{boyorgrl}},
\code{\link{electron}},
\code{\link{hotcore}},
\code{\link{lasers}},
\code{\link{radioact}},
\code{\link{viruses}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
