% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_339.R
\name{ofworth}
\alias{ofworth}
\title{I am a person of worth at least equal to others}
\source{
General Social Survey https://gss.norc.org
}
\description{
ofworth
}
\details{
Question Indicate your agreement with each of the following statements by selecting the number that comes closest to your answer:
D. I feel that I’m a person of worth, at least equal to others.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3039/vshow}{GSS Data Explorer page for \code{ofworth}}.
}

\seealso{
Other Work Organizations: 
\code{\link{afailure}},
\code{\link{nogood}},
\code{\link{notcount}},
\code{\link{optimist}},
\code{\link{pessimst}},
\code{\link{satself}},
\code{\link{slfrspct}}

Other Genes and the Environment: 
\code{\link{afailure}},
\code{\link{genenvo3}},
\code{\link{genenvo4}},
\code{\link{nogood}},
\code{\link{notcount}},
\code{\link{optimist}},
\code{\link{pessimst}},
\code{\link{satself}},
\code{\link{slfrspct}}

Other Feelings: 
\code{\link{afailure}},
\code{\link{nogood}},
\code{\link{notcount}},
\code{\link{optimist}},
\code{\link{pessimst}},
\code{\link{satself}},
\code{\link{slfrspct}}

Other Satisfaction: 
\code{\link{afailure}},
\code{\link{nogood}},
\code{\link{notcount}},
\code{\link{optimist}},
\code{\link{pessimst}},
\code{\link{satself}},
\code{\link{slfrspct}}
}
\concept{Feelings}
\concept{Genes and the Environment}
\concept{Satisfaction}
\concept{Work Organizations}
\keyword{variable}
