% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_583.R
\name{old4}
\alias{old4}
\title{Age of 4th person}
\source{
General Social Survey https://gss.norc.org
}
\description{
old4
}
\details{
Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.)
Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily--on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.)
Are there any people currently staying here--visitors, friends or relatives--do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.)
A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD?
B. Code sex. (ASK IF NOT OBVIOUS).
C. How old was (HEAD/PERSON) on (his/her) last birthday?
D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married?
E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now?
F. Where is (PERSON) staying right now? Is (PERSON) staying at another household;
Is (he/she) traveling; Is (he/she) in some institution or dormitory--like at college or in a hospital or something; or what?
3. Age of fourth person
}
\section{Values}{

\itemize{
\item \code{97} 97 or older
\item \code{98} adult, age unspecified
\item \code{99} child, age unspecified
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5352/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 25 \tab 26 \tab 27 \tab 28 \tab 3 \tab 33 \tab 35 \tab 36 \tab 4 \tab 43 \tab 45 \tab 5 \tab 52 \tab 53 \tab 58 \tab 59 \tab 6 \tab 7 \tab 74 \tab 78 \tab 8 \tab 83 \tab 9 \tab 96 \tab 24 \tab 47 \tab 69 \tab no answer \tab 29 \tab 32 \tab 39 \tab 54 \tab 60 \tab 67 \tab 71 \tab 76 \tab 80 \tab 84 \tab don't know \tab 41 \tab 46 \tab 48 \tab 50 \tab 65 \tab 72 \tab 30 \tab 34 \tab 51 \tab 70 \tab 77 \tab 97 or older \tab 37 \tab 38 \tab 44 \tab 61 \tab 64 \tab 86 \tab 87 \tab 42 \tab 49 \tab 79 \tab 56 \tab 90 \tab 31 \tab 55 \tab 40 \tab 63 \tab 89 \tab 91 \tab 85 \tab 73 \tab 92 \tab 75 \tab 82 \tab 68 \tab 88 \tab 95 \tab 57 \tab 66 \tab 62 \tab 81 \tab adult, age unspecified \tab child, age unspecified \tab skipped on web \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 952 \tab 21 \tab 21 \tab 28 \tab 24 \tab 21 \tab 24 \tab 42 \tab 23 \tab 41 \tab 29 \tab 27 \tab 19 \tab 19 \tab 8 \tab 8 \tab 6 \tab 6 \tab 1 \tab 3 \tab 1 \tab 1 \tab 17 \tab 1 \tab 1 \tab 1 \tab 28 \tab 1 \tab 1 \tab 17 \tab 1 \tab 1 \tab 1 \tab 1 \tab 24 \tab 16 \tab 1 \tab 1 \tab 25 \tab 1 \tab 25 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1255 \tab 15 \tab 9 \tab 14 \tab 15 \tab 15 \tab 11 \tab 15 \tab 16 \tab 8 \tab 9 \tab 6 \tab 5 \tab 16 \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab 2 \tab - \tab - \tab 8 \tab - \tab 1 \tab 1 \tab 11 \tab - \tab - \tab 14 \tab - \tab - \tab - \tab - \tab 11 \tab 15 \tab - \tab - \tab 5 \tab - \tab 11 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1007 \tab 9 \tab 14 \tab 17 \tab 30 \tab 28 \tab 36 \tab 32 \tab 33 \tab 33 \tab 26 \tab 20 \tab 24 \tab 21 \tab 11 \tab 16 \tab 5 \tab 5 \tab - \tab - \tab - \tab - \tab 13 \tab 1 \tab - \tab - \tab 17 \tab - \tab 1 \tab 20 \tab - \tab 1 \tab - \tab - \tab 19 \tab 21 \tab - \tab - \tab 30 \tab 1 \tab 19 \tab - \tab 2 \tab - \tab - \tab 7 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1017 \tab 22 \tab 14 \tab 14 \tab 20 \tab 28 \tab 27 \tab 30 \tab 32 \tab 25 \tab 22 \tab 15 \tab 23 \tab 23 \tab 10 \tab 5 \tab 5 \tab 4 \tab 2 \tab 1 \tab - \tab 1 \tab 22 \tab - \tab - \tab - \tab 25 \tab - \tab - \tab 25 \tab - \tab - \tab - \tab - \tab 30 \tab 14 \tab - \tab - \tab 22 \tab 1 \tab 32 \tab - \tab 3 \tab - \tab - \tab 9 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1040 \tab 25 \tab 16 \tab 18 \tab 19 \tab 21 \tab 14 \tab 20 \tab 23 \tab 24 \tab 14 \tab 18 \tab 17 \tab 22 \tab 5 \tab 6 \tab 7 \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab 18 \tab 1 \tab - \tab 1 \tab 10 \tab - \tab - \tab 18 \tab - \tab - \tab - \tab - \tab 22 \tab 14 \tab - \tab - \tab 25 \tab - \tab 25 \tab - \tab 1 \tab - \tab - \tab 5 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1343 \tab 21 \tab 25 \tab 20 \tab 20 \tab 27 \tab 21 \tab 17 \tab 19 \tab 32 \tab 31 \tab 25 \tab 15 \tab 30 \tab 11 \tab 6 \tab 8 \tab 5 \tab 4 \tab 1 \tab 3 \tab - \tab 29 \tab - \tab 1 \tab 2 \tab 16 \tab - \tab - \tab 21 \tab - \tab 1 \tab - \tab - \tab 18 \tab 20 \tab - \tab 1 \tab 25 \tab - \tab 18 \tab - \tab 4 \tab - \tab - \tab 7 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1104 \tab 17 \tab 22 \tab 21 \tab 16 \tab 24 \tab 25 \tab 24 \tab 26 \tab 14 \tab 16 \tab 21 \tab 9 \tab 31 \tab 12 \tab 12 \tab 4 \tab 11 \tab 9 \tab - \tab 2 \tab 2 \tab 14 \tab - \tab 1 \tab - \tab 24 \tab - \tab - \tab 25 \tab - \tab - \tab - \tab - \tab 27 \tab 22 \tab 2 \tab - \tab 19 \tab - \tab 23 \tab - \tab 5 \tab - \tab - \tab 6 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1062 \tab 20 \tab 20 \tab 11 \tab 10 \tab 25 \tab 25 \tab 18 \tab 15 \tab 16 \tab 19 \tab 14 \tab 8 \tab 22 \tab 8 \tab 11 \tab 10 \tab 6 \tab 2 \tab 1 \tab 1 \tab - \tab 13 \tab - \tab 1 \tab 1 \tab 26 \tab - \tab - \tab 16 \tab - \tab - \tab - \tab - \tab 19 \tab 20 \tab - \tab - \tab 24 \tab - \tab 14 \tab - \tab 7 \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1112 \tab 11 \tab 15 \tab 21 \tab 13 \tab 16 \tab 16 \tab 20 \tab 24 \tab 30 \tab 17 \tab 13 \tab 15 \tab 23 \tab 10 \tab 3 \tab 7 \tab 2 \tab 1 \tab 4 \tab 2 \tab 4 \tab 16 \tab 1 \tab 1 \tab - \tab 24 \tab - \tab 1 \tab 29 \tab - \tab - \tab - \tab - \tab 15 \tab 16 \tab - \tab - \tab 18 \tab - \tab 16 \tab - \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1032 \tab 16 \tab 16 \tab 18 \tab 27 \tab 18 \tab 13 \tab 15 \tab 27 \tab 21 \tab 18 \tab 13 \tab 11 \tab 21 \tab 9 \tab 6 \tab 5 \tab 4 \tab 5 \tab 4 \tab 3 \tab 1 \tab 12 \tab 1 \tab 1 \tab - \tab 22 \tab - \tab - \tab 30 \tab - \tab - \tab - \tab - \tab 13 \tab 27 \tab - \tab 1 \tab 18 \tab - \tab 18 \tab - \tab 7 \tab - \tab - \tab 7 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1282 \tab 11 \tab 21 \tab 23 \tab 28 \tab 21 \tab 20 \tab 33 \tab 18 \tab 23 \tab 24 \tab 15 \tab 15 \tab 21 \tab 17 \tab 13 \tab 12 \tab 5 \tab 2 \tab 3 \tab - \tab 6 \tab 19 \tab 2 \tab 2 \tab - \tab 24 \tab 1 \tab - \tab 25 \tab - \tab - \tab - \tab - \tab 24 \tab 29 \tab - \tab 1 \tab 26 \tab 1 \tab 23 \tab - \tab 6 \tab - \tab - \tab 5 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1097 \tab 19 \tab 17 \tab 27 \tab 23 \tab 15 \tab 20 \tab 11 \tab 11 \tab 9 \tab 13 \tab 7 \tab 7 \tab 19 \tab 6 \tab 6 \tab 8 \tab 2 \tab - \tab 1 \tab 3 \tab 1 \tab 12 \tab - \tab - \tab - \tab 24 \tab - \tab - \tab 26 \tab - \tab 1 \tab - \tab - \tab 19 \tab 16 \tab 1 \tab - \tab 25 \tab - \tab 16 \tab - \tab 3 \tab - \tab - \tab 7 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1098 \tab 13 \tab 22 \tab 21 \tab 17 \tab 20 \tab 17 \tab 25 \tab 19 \tab 20 \tab 18 \tab 11 \tab 13 \tab 21 \tab 10 \tab 15 \tab 9 \tab 4 \tab 1 \tab 3 \tab 5 \tab - \tab 16 \tab - \tab - \tab 1 \tab 14 \tab - \tab - \tab 20 \tab - \tab - \tab - \tab - \tab 19 \tab 17 \tab - \tab 1 \tab 32 \tab - \tab 18 \tab - \tab 2 \tab - \tab - \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1032 \tab 10 \tab 21 \tab 14 \tab 20 \tab 19 \tab 17 \tab 10 \tab 14 \tab 12 \tab 5 \tab 8 \tab 12 \tab 15 \tab 7 \tab 4 \tab 4 \tab 6 \tab 4 \tab 1 \tab 1 \tab 1 \tab 15 \tab - \tab - \tab - \tab 18 \tab - \tab - \tab 19 \tab - \tab - \tab - \tab - \tab 16 \tab 25 \tab - \tab - \tab 16 \tab 2 \tab 9 \tab - \tab 2 \tab - \tab - \tab 7 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1125 \tab 20 \tab 24 \tab 22 \tab 16 \tab 17 \tab 9 \tab 16 \tab 12 \tab 20 \tab 8 \tab 5 \tab 11 \tab 21 \tab 10 \tab 8 \tab 7 \tab 5 \tab 3 \tab 1 \tab 1 \tab 1 \tab 25 \tab 1 \tab 1 \tab - \tab 19 \tab - \tab - \tab 15 \tab - \tab - \tab - \tab - \tab 20 \tab 17 \tab 1 \tab - \tab 15 \tab - \tab 20 \tab - \tab 1 \tab - \tab - \tab 8 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1190 \tab 11 \tab 22 \tab 18 \tab 21 \tab 18 \tab 27 \tab 23 \tab 16 \tab 11 \tab 11 \tab 13 \tab 11 \tab 31 \tab 7 \tab 2 \tab 5 \tab 5 \tab 2 \tab - \tab 3 \tab 2 \tab 21 \tab - \tab 1 \tab - \tab 17 \tab - \tab - \tab 25 \tab - \tab - \tab - \tab - \tab 19 \tab 20 \tab - \tab - \tab 24 \tab - \tab 18 \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2274 \tab 28 \tab 27 \tab 29 \tab 24 \tab 42 \tab 29 \tab 35 \tab 35 \tab 33 \tab 15 \tab 14 \tab 14 \tab 41 \tab 18 \tab 10 \tab 7 \tab 8 \tab 4 \tab 6 \tab 4 \tab 3 \tab 45 \tab 2 \tab 2 \tab - \tab 30 \tab - \tab - \tab 38 \tab - \tab - \tab - \tab - \tab 32 \tab 35 \tab - \tab 1 \tab 34 \tab - \tab 36 \tab - \tab 6 \tab 2 \tab - \tab 6 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 3 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2184 \tab 14 \tab 49 \tab 28 \tab 37 \tab 31 \tab 45 \tab 28 \tab 35 \tab 26 \tab 18 \tab 16 \tab 13 \tab 43 \tab 12 \tab 8 \tab 8 \tab 7 \tab 2 \tab 3 \tab 3 \tab 4 \tab 39 \tab 2 \tab 2 \tab 2 \tab 34 \tab - \tab 1 \tab 27 \tab 1 \tab - \tab - \tab - \tab 39 \tab 38 \tab - \tab 1 \tab 36 \tab - \tab 26 \tab - \tab 8 \tab 1 \tab - \tab 16 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab 1 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2195 \tab - \tab 33 \tab 30 \tab 48 \tab 34 \tab 32 \tab 27 \tab 24 \tab 21 \tab 17 \tab 14 \tab 12 \tab 28 \tab 9 \tab 10 \tab 2 \tab 6 \tab 3 \tab 5 \tab 1 \tab 1 \tab 35 \tab 1 \tab 2 \tab - \tab 33 \tab - \tab - \tab 29 \tab - \tab - \tab - \tab - \tab 34 \tab 35 \tab - \tab - \tab 28 \tab - \tab 29 \tab - \tab 6 \tab - \tab - \tab 28 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 4 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2127 \tab 22 \tab 41 \tab 33 \tab 27 \tab 31 \tab 33 \tab 27 \tab 25 \tab 20 \tab 25 \tab 21 \tab 16 \tab 32 \tab 10 \tab 8 \tab 3 \tab 6 \tab 3 \tab 3 \tab 2 \tab 4 \tab 42 \tab 1 \tab 1 \tab - \tab 36 \tab 1 \tab 1 \tab 42 \tab - \tab 1 \tab - \tab - \tab 38 \tab 29 \tab 2 \tab - \tab 38 \tab - \tab 27 \tab - \tab 5 \tab - \tab - \tab 14 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2266 \tab 1 \tab 36 \tab 24 \tab 20 \tab 26 \tab 23 \tab 17 \tab 19 \tab 18 \tab 12 \tab 15 \tab 10 \tab 25 \tab 14 \tab 11 \tab 11 \tab 5 \tab 2 \tab 2 \tab 4 \tab 1 \tab 30 \tab 2 \tab 1 \tab - \tab 19 \tab 1 \tab - \tab 25 \tab - \tab - \tab 1 \tab - \tab 26 \tab 24 \tab - \tab - \tab 16 \tab - \tab 22 \tab - \tab 4 \tab - \tab 1 \tab 20 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2214 \tab - \tab 35 \tab 28 \tab 27 \tab 32 \tab 35 \tab 30 \tab 25 \tab 15 \tab 20 \tab 27 \tab 17 \tab 47 \tab 8 \tab 17 \tab 7 \tab 7 \tab 2 \tab 5 \tab - \tab 4 \tab 19 \tab - \tab 1 \tab - \tab 23 \tab 1 \tab 1 \tab 25 \tab - \tab - \tab - \tab - \tab 24 \tab 25 \tab - \tab - \tab 26 \tab - \tab 26 \tab - \tab 5 \tab - \tab - \tab 15 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 3500 \tab 32 \tab 38 \tab 54 \tab 48 \tab 46 \tab 56 \tab 53 \tab 42 \tab 20 \tab 30 \tab 39 \tab 24 \tab 45 \tab 30 \tab 10 \tab 11 \tab 9 \tab 10 \tab 7 \tab 7 \tab 1 \tab 39 \tab - \tab - \tab - \tab 54 \tab - \tab 1 \tab 49 \tab - \tab - \tab - \tab - \tab 41 \tab 42 \tab - \tab - \tab 44 \tab - \tab 45 \tab - \tab 7 \tab - \tab - \tab 17 \tab 4 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 23 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 5 \tab 3 \tab 2 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1547 \tab 12 \tab 26 \tab 13 \tab 14 \tab 25 \tab 25 \tab 19 \tab 24 \tab 13 \tab 13 \tab 11 \tab 13 \tab 30 \tab 13 \tab 5 \tab 6 \tab 2 \tab 3 \tab 3 \tab 2 \tab 5 \tab 23 \tab 1 \tab 1 \tab - \tab 21 \tab - \tab - \tab 22 \tab 1 \tab - \tab - \tab - \tab 11 \tab 26 \tab - \tab - \tab 20 \tab - \tab 20 \tab - \tab 6 \tab - \tab - \tab 17 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 15 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1627 \tab 8 \tab 16 \tab 18 \tab 14 \tab 14 \tab 14 \tab 15 \tab 15 \tab 17 \tab 11 \tab 14 \tab 15 \tab 17 \tab 7 \tab 8 \tab 10 \tab 9 \tab 5 \tab 2 \tab 5 \tab 2 \tab 11 \tab 3 \tab 1 \tab 2 \tab 22 \tab 1 \tab - \tab 18 \tab - \tab 1 \tab - \tab - \tab 20 \tab 18 \tab - \tab - \tab 20 \tab - \tab 19 \tab - \tab 4 \tab 1 \tab - \tab 17 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 4 \tab - \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1526 \tab 7 \tab 21 \tab 17 \tab 17 \tab 18 \tab 20 \tab 21 \tab 20 \tab 15 \tab 17 \tab 17 \tab 10 \tab 17 \tab 12 \tab 7 \tab 11 \tab 7 \tab 8 \tab 3 \tab - \tab 1 \tab 17 \tab - \tab - \tab - \tab 20 \tab - \tab - \tab 22 \tab - \tab - \tab 1 \tab - \tab 25 \tab 18 \tab - \tab - \tab 21 \tab - \tab 17 \tab - \tab - \tab 1 \tab - \tab 6 \tab 3 \tab - \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 8 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 7 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2040 \tab 8 \tab 17 \tab 14 \tab 16 \tab 21 \tab 20 \tab 28 \tab 27 \tab 12 \tab 17 \tab 19 \tab 6 \tab 15 \tab 18 \tab 12 \tab 12 \tab 5 \tab 9 \tab 9 \tab 2 \tab 1 \tab 21 \tab 1 \tab 1 \tab 1 \tab 17 \tab 1 \tab 2 \tab 19 \tab 2 \tab - \tab 1 \tab 1 \tab 21 \tab 21 \tab - \tab 2 \tab 19 \tab 1 \tab 18 \tab - \tab 4 \tab - \tab - \tab 38 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2353 \tab 11 \tab 24 \tab 20 \tab 20 \tab 21 \tab 17 \tab 22 \tab 19 \tab 29 \tab 19 \tab 17 \tab 18 \tab 18 \tab 11 \tab 12 \tab 6 \tab 5 \tab 7 \tab 2 \tab 2 \tab 1 \tab 18 \tab 1 \tab 1 \tab 2 \tab 21 \tab - \tab - \tab 16 \tab 1 \tab - \tab - \tab 1 \tab 8 \tab 24 \tab 1 \tab - \tab 18 \tab - \tab 25 \tab - \tab 12 \tab - \tab 1 \tab 23 \tab 4 \tab 1 \tab 1 \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab 8 \tab 1 \tab 1 \tab 2 \tab 3 \tab - \tab - \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 1899 \tab 14 \tab 20 \tab 21 \tab 7 \tab 20 \tab 17 \tab 22 \tab 12 \tab 18 \tab 16 \tab 23 \tab 12 \tab 15 \tab 6 \tab 16 \tab 9 \tab 7 \tab 8 \tab 5 \tab 4 \tab 3 \tab 19 \tab 3 \tab - \tab - \tab 20 \tab 1 \tab - \tab 12 \tab - \tab - \tab 1 \tab 1 \tab 13 \tab 21 \tab 1 \tab - \tab 14 \tab - \tab 14 \tab - \tab 4 \tab - \tab - \tab 13 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 12 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 3 \tab - \tab 3 \tab - \tab - \tab 2 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 3166 \tab 25 \tab 41 \tab 24 \tab 32 \tab 45 \tab 35 \tab 49 \tab 38 \tab 27 \tab 36 \tab 32 \tab 26 \tab 29 \tab 18 \tab 17 \tab 15 \tab 22 \tab 6 \tab 7 \tab 6 \tab 8 \tab 25 \tab 1 \tab 1 \tab 4 \tab 19 \tab - \tab 1 \tab 25 \tab - \tab 1 \tab 1 \tab 3 \tab 21 \tab 38 \tab - \tab 1 \tab 33 \tab 2 \tab 30 \tab - \tab 9 \tab 1 \tab - \tab - \tab 7 \tab 4 \tab 2 \tab - \tab 3 \tab 1 \tab - \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 3 \tab 1 \tab - \tab 7 \tab 6 \tab - \tab - \tab 1 \tab 1 \tab 4 \tab 6 \tab 2 \tab 3 \tab 1 \tab 1 \tab - \tab 1 \tab 2 \tab - \tab 3 \tab - \tab 4 \tab 3 \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab 3 \tab - \tab 2 \tab 2 \tab 7 \tab 12 \tab 3 \tab - \tab 4032 \cr
   2022 \tab 3211 \tab 11 \tab 15 \tab 11 \tab 11 \tab 9 \tab 17 \tab 12 \tab 7 \tab 21 \tab 4 \tab 9 \tab 6 \tab 18 \tab 7 \tab 10 \tab 11 \tab 5 \tab - \tab 5 \tab - \tab - \tab 12 \tab - \tab 3 \tab 1 \tab 14 \tab 1 \tab 1 \tab 13 \tab 1 \tab - \tab - \tab 3 \tab 18 \tab 17 \tab - \tab - \tab 13 \tab - \tab 14 \tab - \tab 6 \tab - \tab 1 \tab 3 \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 6 \tab - \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 56478 \tab 454 \tab 738 \tab 671 \tab 681 \tab 748 \tab 740 \tab 756 \tab 695 \tab 644 \tab 550 \tab 519 \tab 427 \tab 776 \tab 336 \tab 284 \tab 233 \tab 184 \tab 110 \tab 93 \tab 68 \tab 59 \tab 665 \tab 26 \tab 28 \tab 19 \tab 702 \tab 9 \tab 11 \tab 727 \tab 7 \tab 7 \tab 6 \tab 10 \tab 686 \tab 720 \tab 9 \tab 10 \tab 729 \tab 9 \tab 673 \tab 2 \tab 131 \tab 7 \tab 4 \tab 317 \tab 47 \tab 22 \tab 19 \tab 7 \tab 10 \tab 4 \tab 5 \tab 6 \tab 5 \tab 4 \tab 75 \tab 12 \tab 13 \tab 10 \tab 19 \tab 6 \tab 5 \tab 52 \tab 33 \tab 6 \tab 8 \tab 6 \tab 4 \tab 17 \tab 27 \tab 12 \tab 7 \tab 7 \tab 2 \tab 3 \tab 9 \tab 10 \tab 2 \tab 8 \tab 2 \tab 34 \tab 10 \tab 28 \tab 6 \tab 2 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab 3 \tab 4 \tab 1 \tab 9 \tab 1 \tab 5 \tab 2 \tab 9 \tab 18 \tab 3 \tab 3309 \tab 75699 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away3}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{mar3}},
\code{\link{old3}},
\code{\link{relate3}},
\code{\link{relate4}},
\code{\link{where2}},
\code{\link{where3}}

Other Family: 
\code{\link{away3}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{mar3}},
\code{\link{old3}},
\code{\link{relate3}},
\code{\link{relate4}},
\code{\link{where2}},
\code{\link{where3}}

Other Household: 
\code{\link{away3}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{mar3}},
\code{\link{old3}},
\code{\link{relate3}},
\code{\link{relate4}},
\code{\link{where2}},
\code{\link{where3}}
}
\concept{Family}
\concept{Household}
\concept{Household Members and Structure}
\keyword{variable}
