% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_529.R
\name{onenatn}
\alias{onenatn}
\title{One nation or allowed to become fully separate?}
\source{
General Social Survey https://gss.norc.org
}
\description{
onenatn
}
\details{
Question Which of these two statements comes closer to your own view?
}
\section{Values}{

\itemize{
\item \verb{[1]} essential to remain one nation
\item \verb{[2]} allowed to become separate
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4826/vshow}{GSS Data Explorer page for \code{onenatn}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab essential to remain one nation \tab allowed to become separate \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 1155 \tab 69 \tab 120 \tab 23 \tab 2904 \cr
   Total \tab 1537 \tab 1155 \tab 69 \tab 120 \tab 23 \tab 2904 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{ambornin}},
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{clseusa}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{movetown}},
\code{\link{moveusa}}
}
\concept{ISSP National Identity}
\keyword{variable}
