% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_467.R
\name{opeduc}
\alias{opeduc}
\title{Need good education to get ahead}
\source{
General Social Survey https://gss.norc.org
}
\description{
opeduc
}
\details{
Question Please show for each of these how important you think it is for getting ahead in life . . .
C. Having a good education yourself?
}
\section{Values}{

\itemize{
\item \code{1} essential
\item \code{2} very important
\item \code{3} fairly important
\item \code{4} not very important
\item \code{5} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1987 \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4243/vshow}{GSS Data Explorer page for \code{opeduc}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab essential \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1987 \tab 255 \tab 547 \tab 765 \tab 209 \tab 20 \tab 1 \tab 4 \tab 18 \tab - \tab 1819 \cr
   2021 \tab 2029 \tab 623 \tab 875 \tab 313 \tab 32 \tab 10 \tab 21 \tab 116 \tab 13 \tab 4032 \cr
   Total \tab 2284 \tab 1170 \tab 1640 \tab 522 \tab 52 \tab 11 \tab 25 \tab 134 \tab 13 \tab 5851 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{opable}},
\code{\link{opambit}},
\code{\link{opclout}},
\code{\link{ophrdwrk}},
\code{\link{opknow}},
\code{\link{oppared}},
\code{\link{oprace}},
\code{\link{opregion}},
\code{\link{oprelig}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
