% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_125.R
\name{openmind}
\alias{openmind}
\title{Treat dissenters with leniency}
\source{
General Social Survey https://gss.norc.org
}
\description{
openmind
}
\details{
Question Now I am going to read you several statements. For each one please tell me whether you agree strongly, agree, disagree, or disagree strongly.
D. It is best to treat dissenters with leniency and an open mind, since new ideas are the lifeblood of progressive change
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1169/vshow}{GSS Data Explorer page for \code{openmind}}.
}

\seealso{
Other Intergroup Relations: 
\code{\link{ownthing}},
\code{\link{talkback}},
\code{\link{twoclass}}
}
\concept{Intergroup Relations}
\concept{Values}
\keyword{variable}
