% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_125.R
\name{openmind}
\alias{openmind}
\title{Treat dissenters with leniency}
\source{
General Social Survey https://gss.norc.org
}
\description{
openmind
}
\details{
Question Now I am going to read you several statements. For each one please tell me whether you agree strongly, agree, disagree, or disagree strongly.
D. It is best to treat dissenters with leniency and an open mind, since new ideas are the lifeblood of progressive change
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1169/vshow}{GSS Data Explorer page for \code{openmind}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 121 \tab 836 \tab 223 \tab 31 \tab 150 \tab 11 \tab 1372 \cr
   Total \tab 0 \tab 121 \tab 836 \tab 223 \tab 31 \tab 150 \tab 11 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{ownthing}},
\code{\link{talkback}},
\code{\link{twoclass}}
}
\concept{Intergroup Relations}
\keyword{variable}
