% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_430.R
\name{openrel2}
\alias{openrel2}
\title{R and leader ever become an openly acknowledged couple}
\source{
General Social Survey https://gss.norc.org
}
\description{
openrel2
}
\details{
Question Thinking about the leader with whom you had this experience most recently, did you and this leader ever become an openly acknowledged couple?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3940/vshow}{GSS Data Explorer page for \code{openrel2}}.
}

\seealso{
Other Clergy/Sex: 
\code{\link{clrgmar1}},
\code{\link{clrgmar2}},
\code{\link{clrgmar3}},
\code{\link{counsel1}},
\code{\link{counsel2}},
\code{\link{counsel3}},
\code{\link{numclerg}},
\code{\link{openrel1}},
\code{\link{openrel3}}

Other Religion: 
\code{\link{clrgmar1}},
\code{\link{clrgmar2}},
\code{\link{clrgmar3}},
\code{\link{counsel1}},
\code{\link{counsel2}},
\code{\link{counsel3}},
\code{\link{numclerg}},
\code{\link{openrel1}},
\code{\link{openrel3}}

Other Romantic Relationships: 
\code{\link{openrel1}},
\code{\link{openrel3}}

Other Sex: 
\code{\link{clrgmar1}},
\code{\link{clrgmar2}},
\code{\link{clrgmar3}},
\code{\link{counsel1}},
\code{\link{counsel2}},
\code{\link{counsel3}},
\code{\link{numclerg}},
\code{\link{openrel1}},
\code{\link{openrel3}}
}
\concept{Clergy/Sex}
\concept{Religion}
\concept{Romantic Relationships}
\concept{Sex}
\keyword{variable}
