% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_437.R
\name{opfamily}
\alias{opfamily}
\title{Family bkgd determines success in u.s.}
\source{
General Social Survey https://gss.norc.org
}
\description{
opfamily
}
\details{
Question Please indicate whether you agree or disagree with each of the following statements.
C. In America what you achieve in life depends largely on your family background.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4008/vshow}{GSS Data Explorer page for \code{opfamily}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab 5 \tab no answer \tab Total \cr
   1985 \tab 857 \tab 64 \tab 138 \tab 123 \tab 274 \tab 62 \tab 16 \tab 1534 \cr
   Total \tab 857 \tab 64 \tab 138 \tab 123 \tab 274 \tab 62 \tab 16 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{fecolop}},
\code{\link{fehlpbus}},
\code{\link{fehlpcol}},
\code{\link{fehlpjob}},
\code{\link{feinc}},
\code{\link{fejobop}},
\code{\link{hsbasics}},
\code{\link{hsrespct}},
\code{\link{hssexed}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
