% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_467.R
\name{opknow}
\alias{opknow}
\title{Need to know right people to get ahead}
\source{
General Social Survey https://gss.norc.org
}
\description{
opknow
}
\details{
Question Please show for each of these how important you think it is for getting ahead in life . . .
G. Knowing the right people?
}
\section{Values}{

\itemize{
\item \code{1} essential
\item \code{2} very important
\item \code{3} fairly important
\item \code{4} not very important
\item \code{5} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab -/-/- \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2021 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4247/vshow}{GSS Data Explorer page for \code{opknow}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{opable}},
\code{\link{opambit}},
\code{\link{opclout}},
\code{\link{opeduc}},
\code{\link{ophrdwrk}},
\code{\link{oppared}},
\code{\link{oprace}},
\code{\link{opregion}},
\code{\link{oprelig}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
