% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_147.R
\name{opoutcme}
\alias{opoutcme}
\title{Should amer promote eq oppty or eq outcomes?}
\source{
General Social Survey https://gss.norc.org
}
\description{
opoutcme
}
\details{
Question Some people think America should promote equal opportunity for all, that is, allowing everyone to compete for jobs  and wealth on a fair and even basis. Other people think America should promote equal outcomes, that is, insuring that everyone has a decent standard of living and that there are only small differences in wealth and income betweenthe top and bottom in society.
Which do you favor:  promoting equal opportunity or promoting equal outcomes?
}
\section{Values}{

\itemize{
\item \code{1} promote equal oppotunity
\item \code{2} promote equal outcomes
\item \code{3} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1379/vshow}{GSS Data Explorer page for \code{opoutcme}}.
}

\seealso{
Other Culture: 
\code{\link{bigband}},
\code{\link{blugrass}},
\code{\link{lfechnce}},
\code{\link{lfegenes}},
\code{\link{lfegod}},
\code{\link{lfehrdwk}},
\code{\link{lfesocty}},
\code{\link{obtohelp}},
\code{\link{united}}
}
\concept{Culture}
\concept{Inequality, Income}
\keyword{variable}
