% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_147.R
\name{opoutcme}
\alias{opoutcme}
\title{Should amer promote eq oppty or eq outcomes?}
\source{
General Social Survey https://gss.norc.org
}
\description{
opoutcme
}
\details{
Question Some people think America should promote equal opportunity for all, that is, allowing everyone to compete for jobs  and wealth on a fair and even basis. Other people think America should promote equal outcomes, that is, insuring that everyone has a decent standard of living and that there are only small differences in wealth and income betweenthe top and bottom in society.
Which do you favor:  promoting equal opportunity or promoting equal outcomes?
}
\section{Values}{

\itemize{
\item \verb{[1]} promote equal oppotunity
\item \verb{[2]} promote equal outcomes
\item \verb{[3]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1379/vshow}{GSS Data Explorer page for \code{opoutcme}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab promote equal oppotunity \tab promote equal outcomes \tab other \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 1347 \tab 194 \tab 17 \tab 41 \tab 7 \tab 1606 \cr
   Total \tab 0 \tab 1347 \tab 194 \tab 17 \tab 41 \tab 7 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{bigband}},
\code{\link{blugrass}},
\code{\link{lfechnce}},
\code{\link{lfegenes}},
\code{\link{lfegod}},
\code{\link{lfehrdwk}},
\code{\link{lfesocty}},
\code{\link{obtohelp}},
\code{\link{united}}
}
\concept{Culture}
\keyword{variable}
