% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_468.R
\name{oppol}
\alias{oppol}
\title{Must have certain polviews to get ahead}
\source{
General Social Survey https://gss.norc.org
}
\description{
oppol
}
\details{
Question Please show for each of these how important you think it is for getting ahead in life . . .
M. A person''s political beliefs?
}
\section{Values}{

\itemize{
\item \code{1} essential
\item \code{2} very important
\item \code{3} fairly important
\item \code{4} not very important
\item \code{5} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4253/vshow}{GSS Data Explorer page for \code{oppol}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{goodlife}},
\code{\link{incentiv}},
\code{\link{inequal1}},
\code{\link{inequal2}},
\code{\link{inequal3}},
\code{\link{inequal4}},
\code{\link{inequal5}},
\code{\link{opsex}}

Other Social Mobility: 
\code{\link{goodlife}},
\code{\link{opsex}}
}
\concept{ISSP Social Inequality}
\concept{Social Mobility}
\keyword{variable}
