% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_541.R
\name{oppsegov}
\alias{oppsegov}
\title{How important: citizen engage in acts of civil disobeyed}
\source{
General Social Survey https://gss.norc.org
}
\description{
oppsegov
}
\details{
Question There are different opinions about people's rights in a democracy. On a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it:
F. That citizens may engage in acts of civil disobedience when they oppose government actions.
}
\section{Values}{

\itemize{
\item \code{1} not at all important
\item \code{7} very important
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4938/vshow}{GSS Data Explorer page for \code{oppsegov}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab not at all important \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 307 \tab 144 \tab 131 \tab 209 \tab 220 \tab 131 \tab 281 \tab 38 \tab 11 \tab 2812 \cr
   2014 \tab 1274 \tab 207 \tab 81 \tab 73 \tab 214 \tab 200 \tab 137 \tab 222 \tab 128 \tab 2 \tab 2538 \cr
   Total \tab 2614 \tab 514 \tab 225 \tab 204 \tab 423 \tab 420 \tab 268 \tab 503 \tab 166 \tab 13 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actlaw}},
\code{\link{affctlaw}},
\code{\link{discpol}},
\code{\link{govdook}},
\code{\link{poleff18}},
\code{\link{poleff19}},
\code{\link{poleff20}},
\code{\link{polgreed}},
\code{\link{polint1}}
}
\concept{ISSP Citizenship}
\keyword{variable}
