% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_468.R
\name{opsex}
\alias{opsex}
\title{Need to be right sex to get ahead}
\source{
General Social Survey https://gss.norc.org
}
\description{
opsex
}
\details{
Question Please show for each of these how important you think it is for getting ahead in life . . .
L. Being born a man or a woman -- how important is that for getting ahead in life?
}
\section{Values}{

\itemize{
\item \code{1} essential
\item \code{2} very important
\item \code{3} fairly important
\item \code{4} not very important
\item \code{5} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1987 \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4252/vshow}{GSS Data Explorer page for \code{opsex}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab essential \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1987 \tab 255 \tab 50 \tab 200 \tab 342 \tab 485 \tab 417 \tab 42 \tab 28 \tab - \tab 1819 \cr
   2021 \tab 2029 \tab 69 \tab 198 \tab 403 \tab 422 \tab 701 \tab 76 \tab 120 \tab 14 \tab 4032 \cr
   Total \tab 2284 \tab 119 \tab 398 \tab 745 \tab 907 \tab 1118 \tab 118 \tab 148 \tab 14 \tab 5851 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{goodlife}},
\code{\link{incentiv}},
\code{\link{inequal1}},
\code{\link{inequal2}},
\code{\link{inequal3}},
\code{\link{inequal4}},
\code{\link{inequal5}},
\code{\link{oppol}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
