% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_466.R
\name{opwlth}
\alias{opwlth}
\title{Need wealthy family to get ahead}
\source{
General Social Survey https://gss.norc.org
}
\description{
opwlth
}
\details{
Question Please show for each of these how important you think it is for getting ahead in life . . .
A. First, how important is coming from a wealthy family?
}
\section{Values}{

\itemize{
\item \verb{[1]} essential
\item \verb{[2]} very important
\item \verb{[3]} fairly important
\item \verb{[4]} not very important
\item \verb{[5]} not important at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   B/C/- \tab 2021 \cr
   Full \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4241/vshow}{GSS Data Explorer page for \code{opwlth}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab essential \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1987 \tab 255 \tab 62 \tab 290 \tab 432 \tab 466 \tab 264 \tab 35 \tab 15 \tab - \tab 1819 \cr
   2000 \tab 1545 \tab 35 \tab 206 \tab 338 \tab 418 \tab 179 \tab 42 \tab 54 \tab - \tab 2817 \cr
   2021 \tab 2029 \tab 118 \tab 407 \tab 580 \tab 506 \tab 204 \tab 63 \tab 117 \tab 8 \tab 4032 \cr
   Total \tab 3829 \tab 215 \tab 903 \tab 1350 \tab 1390 \tab 647 \tab 140 \tab 186 \tab 8 \tab 8668 \cr
}
}

\concept{ISSP Social Inequality}
\keyword{variable}
