% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_198.R
\name{organsw}
\alias{organsw}
\title{Which of these procedures would be the worst?}
\source{
General Social Survey https://gss.norc.org
}
\description{
organsw
}
\details{
Question Recently, medical science has made it possible to save lives by transplanting body parts from donors to patients who need them. But for some body organs there are not enough to go around, and some patients die before they can obtain a transplant. When only one organ is available and several patients need it for survival, the organ could be assigned to a patient by one of the following procedures: (READ PROCEDURES ON SHOWCARD)
By auction: The organ is assigned to the patient who can pay the most for it.
By first come, first served: the organ is assigned to the patient who has been waiting the longest.
By lottery: The organ is assigned to the patient whose name is drawn at random.
By merit: The organ is assigned to the patient who can make the greatest contribution to others and society.
B. Which of these procedures would be the worst way to assign an organ? (CODE ONE RESPONSE ONLY)
}
\section{Values}{

\itemize{
\item \verb{[1]} auction
\item \verb{[2]} first come-first served
\item \verb{[3]} lottery
\item \verb{[4]} merit
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1867/vshow}{GSS Data Explorer page for \code{organsw}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab auction \tab first come-first served \tab lottery \tab merit \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 1102 \tab 16 \tab 110 \tab 171 \tab 37 \tab 8 \tab 2904 \cr
   Total \tab 1460 \tab 1102 \tab 16 \tab 110 \tab 171 \tab 37 \tab 8 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{adoption}},
\code{\link{frndawk}},
\code{\link{frnddeal}},
\code{\link{organsb}},
\code{\link{sellbaby}},
\code{\link{sellorgn}},
\code{\link{sellsex}},
\code{\link{tablprce}},
\code{\link{transrel}}
}
\concept{Markets}
\keyword{variable}
