% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_136.R
\name{orgfin}
\alias{orgfin}
\title{R's financial situation after org. changes}
\source{
General Social Survey https://gss.norc.org
}
\description{
orgfin
}
\details{
Question As a result of these organizational changes, would you say that your financial situation is now much better, better, worse, much worse, or have these things not affected your finances?
}
\section{Values}{

\itemize{
\item \verb{[1]} much better
\item \verb{[2]} better
\item \verb{[3]} worse
\item \verb{[4]} much worse
\item \verb{[5]} no effect
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1270/vshow}{GSS Data Explorer page for \code{orgfin}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab much better \tab better \tab worse \tab much worse \tab no effect \tab don't know \tab no answer \tab Total \cr
   1991 \tab 1045 \tab 19 \tab 58 \tab 79 \tab 37 \tab 247 \tab 1 \tab 31 \tab 1517 \cr
   Total \tab 1045 \tab 19 \tab 58 \tab 79 \tab 37 \tab 247 \tab 1 \tab 31 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{empyears}},
\code{\link{evunemp}},
\code{\link{fndjob1}},
\code{\link{fndjob2}},
\code{\link{fndjob3}},
\code{\link{fndjob4}},
\code{\link{numunemp}},
\code{\link{totunemp}},
\code{\link{wrkyears}}
}
\concept{Work Organizations}
\keyword{variable}
