% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_644.R
\name{orginc}
\alias{orginc}
\title{Financial distribution in household}
\source{
General Social Survey https://gss.norc.org
}
\description{
orginc
}
\details{
Question How do you and your (spousepartfill) organize the income that one or both of you receive? Please choose the option that comes closest.
}
\section{Values}{

\itemize{
\item \code{1} r manage all money and gives spouse/partner his/her share
\item \code{2} r spouse/partner manages all money and gives r share
\item \code{3} r and spouse/partner pool all money and take what we need
\item \code{4} r and spouse/partner pool some money and keep rest separate
\item \code{5} r and spouse/partner keep money separate
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7774/vshow}{GSS Data Explorer page for \code{orginc}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab r manage all money and gives spouse/partner his/her share \tab r spouse/partner manages all money and gives r share \tab r and spouse/partner pool all money and take what we need \tab r and spouse/partner pool some money and keep rest separate \tab r and spouse/partner keep money separate \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 3023 \tab 47 \tab 44 \tab 257 \tab 69 \tab 89 \tab 11 \tab 2 \tab 2 \tab 3544 \cr
   Total \tab 3023 \tab 47 \tab 44 \tab 257 \tab 69 \tab 89 \tab 11 \tab 2 \tab 2 \tab 3544 \cr
}
}

\keyword{variable}
