% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_214.R
\name{ormntldc}
\alias{ormntldc}
\title{In what order go to a psychiatrist for help}
\source{
General Social Survey https://gss.norc.org
}
\description{
ormntldc
}
\details{
Question Here are the things you said NAME should do. INTERVIEWER: READ "YES" ANSWERS AND ASK: In what order should s/he do them?
D. Go to a psychiatrist for help.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2024/vshow}{GSS Data Explorer page for \code{ormntldc}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1914 \tab 75 \tab 135 \tab 167 \tab 169 \tab 154 \tab 98 \tab 48 \tab 16 \tab 2 \tab 4 \tab 122 \tab 2904 \cr
   2006 \tab 3383 \tab 92 \tab 136 \tab 225 \tab 264 \tab 165 \tab 123 \tab 60 \tab 20 \tab 4 \tab - \tab 38 \tab 4510 \cr
   Total \tab 5297 \tab 167 \tab 271 \tab 392 \tab 433 \tab 319 \tab 221 \tab 108 \tab 36 \tab 6 \tab 4 \tab 160 \tab 7414 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{orhealer}},
\code{\link{ormeddoc}},
\code{\link{ormntlhs}},
\code{\link{ormntlot}},
\code{\link{orotcmed}},
\code{\link{orrxmed}},
\code{\link{orslfhlp}},
\code{\link{ortlkclr}},
\code{\link{ortlkfm}}
}
\concept{Mental Health}
\keyword{variable}
