% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_214.R
\name{orotcmed}
\alias{orotcmed}
\title{In what order take non-prescription medication}
\source{
General Social Survey https://gss.norc.org
}
\description{
orotcmed
}
\details{
Question Here are the things you said NAME should do. INTERVIEWER: READ "YES" ANSWERS AND ASK: In what order should s/he do them?
H. Take non-prescription medication, like over the counter sleeping pills.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2028/vshow}{GSS Data Explorer page for \code{orotcmed}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab no answer \tab Total \cr
   1996 \tab 2762 \tab 4 \tab 5 \tab 9 \tab 4 \tab 15 \tab 8 \tab 13 \tab 12 \tab 16 \tab 8 \tab 48 \tab 2904 \cr
   2006 \tab 4315 \tab 3 \tab 4 \tab 8 \tab 11 \tab 25 \tab 13 \tab 34 \tab 28 \tab 33 \tab 10 \tab 26 \tab 4510 \cr
   Total \tab 7077 \tab 7 \tab 9 \tab 17 \tab 15 \tab 40 \tab 21 \tab 47 \tab 40 \tab 49 \tab 18 \tab 74 \tab 7414 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{orhealer}},
\code{\link{ormeddoc}},
\code{\link{ormntldc}},
\code{\link{ormntlhs}},
\code{\link{ormntlot}},
\code{\link{orrxmed}},
\code{\link{orslfhlp}},
\code{\link{ortlkclr}},
\code{\link{ortlkfm}}
}
\concept{Mental Health}
\keyword{variable}
