% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_214.R
\name{orrxmed}
\alias{orrxmed}
\title{In what order take prescription medication}
\source{
General Social Survey https://gss.norc.org
}
\description{
orrxmed
}
\details{
Question Here are the things you said NAME should do. INTERVIEWER: READ "YES" ANSWERS AND ASK: In what order should s/he do them?
I. Take prescription medication.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2029/vshow}{GSS Data Explorer page for \code{orrxmed}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2082 \tab 2 \tab 1 \tab 34 \tab 55 \tab 104 \tab 110 \tab 150 \tab 159 \tab 72 \tab 20 \tab 3 \tab 112 \tab 2904 \cr
   2006 \tab 3489 \tab 8 \tab - \tab 57 \tab 108 \tab 132 \tab 163 \tab 173 \tab 187 \tab 123 \tab 31 \tab - \tab 39 \tab 4510 \cr
   Total \tab 5571 \tab 10 \tab 1 \tab 91 \tab 163 \tab 236 \tab 273 \tab 323 \tab 346 \tab 195 \tab 51 \tab 3 \tab 151 \tab 7414 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{orhealer}},
\code{\link{ormeddoc}},
\code{\link{ormntldc}},
\code{\link{ormntlhs}},
\code{\link{ormntlot}},
\code{\link{orotcmed}},
\code{\link{orslfhlp}},
\code{\link{ortlkclr}},
\code{\link{ortlkfm}}
}
\concept{Mental Health}
\keyword{variable}
