% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_253.R
\name{othbest}
\alias{othbest}
\title{What language other than english does r speak b}
\source{
General Social Survey https://gss.norc.org
}
\description{
othbest
}
\details{
Question D. Which of these languages other than English do you speak most fluently?
}
\section{Values}{

\itemize{
\item \code{1} english
\item \code{2} spanish
\item \code{3} polish
\item \code{4} french
\item \code{5} hebrew
\item \code{6} russian
\item \code{7} tagalog
\item \code{8} chinese
\item \code{9} urdu
\item \code{10} italian
\item \code{11} romanian
\item \code{12} german
\item \code{13} pashtu
\item \code{14} chamorro
\item \code{15} ukranian
\item \code{16} hungarian
\item \code{17} turkish
\item \code{18} khmer
\item \code{19} korean
\item \code{20} navajo
\item \code{21} malayam
\item \code{22} arabic
\item \code{23} farsi-persian
\item \code{24} serbian
\item \code{25} lithuanian
\item \code{27} thai
\item \code{28} gujarati
\item \code{30} dari
\item \code{31} slovak
\item \code{32} japanese
\item \code{33} portuguese
\item \code{34} vajun
\item \code{35} yidddish
\item \code{36} creole
\item \code{37} vietnamese
\item \code{38} norwegian
\item \code{39} malaysian
\item \code{40} latin
\item \code{41} czech
\item \code{42} hindu
\item \code{43} hakka
\item \code{44} swedish
\item \code{45} vesaya
\item \code{46} sign language
\item \code{47} tongan
\item \code{48} armenian
\item \code{49} ibo
\item \code{50} dutch
\item \code{51} cherokee
\item \code{52} samoan
\item \code{53} yoruba
\item \code{54} greek
\item \code{55} lakota
\item \code{56} african
\item \code{57} anishnabee
\item \code{58} ilokand
\item \code{59} assyrian
\item \code{60} danish
\item \code{61} laotian
\item \code{62} finnish
\item \code{63} mohawk
\item \code{64} amharic
\item \code{65} bangladesh
\item \code{66} oriya
\item \code{67} hawaiian
\item \code{69} hmong
\item \code{70} indonesian
\item \code{71} swahili
\item \code{72} telugu
\item \code{73} kannada
\item \code{74} bulgarian
\item \code{75} guarani
\item \code{76} bosnian
\item \code{77} punjabi
\item \code{78} latavain
\item \code{79} basque
\item \code{80} bicol
\item \code{81} croatian
\item \code{82} tigan gypsy
\item \code{83} slovenian
\item \code{84} belgian
\item \code{85} mandingo
\item \code{86} jamaca
\item \code{87} somali
\item \code{88} native amerian
\item \code{89} tagringa
\item \code{90} krio
\item \code{91} indian
\item \code{92} cebuano
\item \code{93} macedonian
\item \code{94} patois
\item \code{95} pangasinan
\item \code{96} ilocano
\item \code{97} marathi
\item \code{98} kashmiri/cashmiri
\item \code{99} tamil
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2286/vshow}{GSS Data Explorer page for \code{othbest}}.
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{asnchng}},
\code{\link{asncls}},
\code{\link{asncom}},
\code{\link{asnrel}},
\code{\link{asnschl}},
\code{\link{asnwrk}},
\code{\link{betrlang}},
\code{\link{blkchng}},
\code{\link{blkcls}},
\code{\link{blkcom}},
\code{\link{blkrel}},
\code{\link{blkschl}},
\code{\link{blkwrk}},
\code{\link{comamind}},
\code{\link{comasn}},
\code{\link{comblk}},
\code{\link{comhisp}},
\code{\link{comjews}},
\code{\link{comwht}},
\code{\link{contblk}},
\code{\link{contchn}},
\code{\link{contcuba}},
\code{\link{conteng}},
\code{\link{contirsh}},
\code{\link{contitl}},
\code{\link{contjew}},
\code{\link{contjpn}},
\code{\link{contmex}},
\code{\link{contmslm}},
\code{\link{contpr}},
\code{\link{contvn}},
\code{\link{engoff1}},
\code{\link{engthrtn}},
\code{\link{engunite}},
\code{\link{engvote}},
\code{\link{ethchng}},
\code{\link{fairasns}},
\code{\link{fairblks}},
\code{\link{fairhsps}},
\code{\link{fairjews}},
\code{\link{fairwhts}},
\code{\link{famasns}},
\code{\link{famblks}},
\code{\link{famhsps}},
\code{\link{famjews}},
\code{\link{famwhts}},
\code{\link{forlang1}},
\code{\link{getlang}},
\code{\link{granlang}},
\code{\link{hispchng}},
\code{\link{hispcls}},
\code{\link{hispcom}},
\code{\link{hisprel}},
\code{\link{hispschl}},
\code{\link{hispwork}},
\code{\link{hispwrk}},
\code{\link{hseth}},
\code{\link{hswht}},
\code{\link{immcrmup}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{jewschng}},
\code{\link{jewscls}},
\code{\link{jewscom}},
\code{\link{jewsrel}},
\code{\link{jewsschl}},
\code{\link{jewswrk}},
\code{\link{knwasn}},
\code{\link{knwblk}},
\code{\link{knwhisp}},
\code{\link{knwjew}},
\code{\link{knwwht}},
\code{\link{langcom}},
\code{\link{langwrk}},
\code{\link{leastcom}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{mostcom}},
\code{\link{neieth1}},
\code{\link{neieth10}},
\code{\link{neieth11}},
\code{\link{neieth12}},
\code{\link{neieth13}},
\code{\link{neieth14}},
\code{\link{neieth2}},
\code{\link{neieth3}},
\code{\link{neieth4}},
\code{\link{neieth5}},
\code{\link{neieth6}},
\code{\link{neieth7}},
\code{\link{neieth8}},
\code{\link{neieth9}},
\code{\link{nobiling}},
\code{\link{othhome}},
\code{\link{othlang}},
\code{\link{othlang1}},
\code{\link{othlang2}},
\code{\link{parlang}},
\code{\link{spklang}},
\code{\link{spklangw}},
\code{\link{twolang}},
\code{\link{usamind}},
\code{\link{usasn}},
\code{\link{usblk}},
\code{\link{uselang}},
\code{\link{ushisp}},
\code{\link{usjews}},
\code{\link{usmixed}},
\code{\link{uswht}},
\code{\link{whtchng}},
\code{\link{whtcls}},
\code{\link{whtcom}},
\code{\link{whtrel}},
\code{\link{whtschl}},
\code{\link{whtwrk}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
