% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_141.R
\name{othduty2}
\alias{othduty2}
\title{You train employees in other ways, not supervise}
\source{
General Social Survey https://gss.norc.org
}
\description{
othduty2
}
\details{
Question (On the card you have are/Here is a card listing) 7 different kinds of work activities. Which ones do you do with employees whom you do not officially supervise?  Please mention all that apply:
B. Train employees in other ways.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1326/vshow}{GSS Data Explorer page for \code{othduty2}}.
}

\seealso{
Other Work Organizations: 
\code{\link{othduty1}},
\code{\link{othduty3}},
\code{\link{othduty4}},
\code{\link{supduty2}},
\code{\link{supduty3}},
\code{\link{supduty4}},
\code{\link{supduty5}},
\code{\link{supduty6}},
\code{\link{supduty7}}
}
\concept{Work Organizations}
\keyword{variable}
