% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_142.R
\name{othduty7}
\alias{othduty7}
\title{You promote or discipline actions, not supervise}
\source{
General Social Survey https://gss.norc.org
}
\description{
othduty7
}
\details{
Question (On the card you have are/Here is a card listing) 7 different kinds of work activities. Which ones do you do with employees whom you do not officially supervise?  Please mention all that apply:
G. Recommend promotions or disciplinary actions.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1331/vshow}{GSS Data Explorer page for \code{othduty7}}.
}

\seealso{
Other Work Organizations: 
\code{\link{jobjudge}},
\code{\link{orgmoney}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{quality}},
\code{\link{quantity}},
\code{\link{totmoney}},
\code{\link{wrkwell}},
\code{\link{youmoney}}
}
\concept{Work Organizations}
\keyword{variable}
