% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_508.R
\name{otherwrk}
\alias{otherwrk}
\title{Does r have any additional job}
\source{
General Social Survey https://gss.norc.org
}
\description{
otherwrk
}
\details{
Question In addition to your main job, do you do any other work for additional income?
FI INSTRUCTION: PROBE ‘YES’ RESPONSES
}
\section{Values}{

\itemize{
\item \code{1} no
\item \code{2} yes, mostly as an employee
\item \code{3} yes, mostly on a self-employed basis
\item \code{4} yes, other
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4624/vshow}{GSS Data Explorer page for \code{otherwrk}}.
}

\seealso{
Other Work Orientation: 
\code{\link{nounemp3}},
\code{\link{nounemp4}},
\code{\link{skltrain}},
\code{\link{spwrkgvt}}
}
\concept{Work Orientation}
\keyword{variable}
