% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_98.R
\name{othfam4}
\alias{othfam4}
\title{Person #4 is r's family member}
\source{
General Social Survey https://gss.norc.org
}
\description{
othfam4
}
\details{
Question Here is a list of some of the ways in which people are connected to each other. Some people can be connected to you In more than one way. For example, a man could be your brother and he may belong to your church and be your lawyer.
When I read you a name, please tell me all the ways that person is connected to you.
How is (NAME) connected to you?  PROBE:  What other ways?  ASK FOR EACH NAME.
E4. Fourth Person is Other Family Member
}
\section{Values}{

\itemize{
\item \code{1} mentioned
\item \code{2} not mentioned
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/898/vshow}{GSS Data Explorer page for \code{othfam4}}.
}

\seealso{
Other Social Networks: 
\code{\link{advisor1}},
\code{\link{advisor2}},
\code{\link{advisor3}},
\code{\link{advisor4}},
\code{\link{advisor5}},
\code{\link{age1}},
\code{\link{age2}},
\code{\link{age3}},
\code{\link{age4}},
\code{\link{age5}},
\code{\link{child1}},
\code{\link{child2}},
\code{\link{child3}},
\code{\link{child4}},
\code{\link{child5}},
\code{\link{close12}},
\code{\link{close13}},
\code{\link{close14}},
\code{\link{close15}},
\code{\link{close23}},
\code{\link{close24}},
\code{\link{close25}},
\code{\link{close34}},
\code{\link{close35}},
\code{\link{close45}},
\code{\link{cowork1}},
\code{\link{cowork2}},
\code{\link{cowork3}},
\code{\link{cowork4}},
\code{\link{cowork5}},
\code{\link{educ1}},
\code{\link{educ2}},
\code{\link{educ3}},
\code{\link{educ4}},
\code{\link{educ5}},
\code{\link{eqclose}},
\code{\link{friend1}},
\code{\link{friend2}},
\code{\link{friend3}},
\code{\link{friend4}},
\code{\link{friend5}},
\code{\link{frndknow}},
\code{\link{known1}},
\code{\link{known2}},
\code{\link{known3}},
\code{\link{known4}},
\code{\link{known5}},
\code{\link{memgrp1}},
\code{\link{memgrp2}},
\code{\link{memgrp3}},
\code{\link{memgrp4}},
\code{\link{memgrp5}},
\code{\link{neighbr1}},
\code{\link{neighbr2}},
\code{\link{neighbr3}},
\code{\link{neighbr4}},
\code{\link{neighbr5}},
\code{\link{numgiven}},
\code{\link{other1}},
\code{\link{other2}},
\code{\link{other3}},
\code{\link{other4}},
\code{\link{other5}},
\code{\link{othfam1}},
\code{\link{othfam2}},
\code{\link{othfam3}},
\code{\link{othfam5}},
\code{\link{parent1}},
\code{\link{parent2}},
\code{\link{parent3}},
\code{\link{parent4}},
\code{\link{parent5}},
\code{\link{partyid1}},
\code{\link{partyid2}},
\code{\link{partyid3}},
\code{\link{race1}},
\code{\link{race2}},
\code{\link{race3}},
\code{\link{race4}},
\code{\link{race5}},
\code{\link{rclose1}},
\code{\link{rclose2}},
\code{\link{rclose3}},
\code{\link{rclose4}},
\code{\link{rclose5}},
\code{\link{relig1}},
\code{\link{relig2}},
\code{\link{relig3}},
\code{\link{relig4}},
\code{\link{relig5}},
\code{\link{sex1}},
\code{\link{sex2}},
\code{\link{sex3}},
\code{\link{sex4}},
\code{\link{sex5}},
\code{\link{sibling1}},
\code{\link{sibling2}},
\code{\link{sibling3}},
\code{\link{sibling4}},
\code{\link{sibling5}},
\code{\link{spouse1}},
\code{\link{spouse2}},
\code{\link{spouse3}},
\code{\link{spouse4}},
\code{\link{spouse5}},
\code{\link{talkpol}},
\code{\link{talkpol1}},
\code{\link{talkpol2}},
\code{\link{talkpol3}},
\code{\link{talkto1}},
\code{\link{talkto2}},
\code{\link{talkto3}},
\code{\link{talkto4}},
\code{\link{talkto5}}
}
\concept{Social Networks}
\keyword{variable}
