% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_502.R
\name{othhrs}
\alias{othhrs}
\title{Hours r spends at other job(s)}
\source{
General Social Survey https://gss.norc.org
}
\description{
othhrs
}
\details{
Question How many hours per week do you normally work in these other jobs, not including time spent on traveling to work?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4565/vshow}{GSS Data Explorer page for \code{othhrs}}.
}

\seealso{
Other Work Orientation: 
\code{\link{employer}},
\code{\link{localnum}},
\code{\link{moonlite}},
\code{\link{numemply}},
\code{\link{numsites}},
\code{\link{supervis}},
\code{\link{supnum}},
\code{\link{suprvsjb}},
\code{\link{totalnum}}

Other Job: 
\code{\link{localnum}},
\code{\link{moonlite}},
\code{\link{numsites}},
\code{\link{supervis}},
\code{\link{supnum}},
\code{\link{totalnum}}
}
\concept{Job}
\concept{Work Orientation}
\keyword{variable}
