% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_34.R
\name{othjew}
\alias{othjew}
\title{Consider self to be jewish}
\source{
General Social Survey https://gss.norc.org
}
\description{
othjew
}
\details{
Question A. Do you consider yourself to be Jewish for any reason?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} half/part
\item \verb{[3]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/290/vshow}{GSS Data Explorer page for \code{othjew}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab half/part \tab no \tab no answer \tab Total \cr
   2008 \tab 1682 \tab 4 \tab 5 \tab 320 \tab 12 \tab 2023 \cr
   2012 \tab 1579 \tab 7 \tab 5 \tab 375 \tab 8 \tab 1974 \cr
   Total \tab 3261 \tab 11 \tab 10 \tab 695 \tab 20 \tab 3997 \cr
}
}

\seealso{
Other Core: 
\code{\link{denom}},
\code{\link{grass}},
\code{\link{grassy}},
\code{\link{lawfirm}},
\code{\link{lawimp}},
\code{\link{lawinfo}},
\code{\link{other}},
\code{\link{relig}},
\code{\link{wirtap}}
}
\concept{Core}
\keyword{variable}
