% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_40.R
\name{othjew16}
\alias{othjew16}
\title{R raised as a jew}
\source{
General Social Survey https://gss.norc.org
}
\description{
othjew16
}
\details{
Question Were you raised as a Jew?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} half/part
\item \code{3} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/344/vshow}{GSS Data Explorer page for \code{othjew16}}.
}

\seealso{
Other Core: 
\code{\link{denom16}},
\code{\link{fund16}},
\code{\link{jew16}},
\code{\link{jew16aj}},
\code{\link{oth16}},
\code{\link{relig16}},
\code{\link{spden}},
\code{\link{spother}},
\code{\link{sprel}}
}
\concept{Core}
\keyword{variable}
