% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_253.R
\name{othlang}
\alias{othlang}
\title{Can r speak language other than english/spanish}
\source{
General Social Survey https://gss.norc.org
}
\description{
othlang
}
\details{
Question A. Can you speak a language other than English (Spanish)?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2283/vshow}{GSS Data Explorer page for \code{othlang}}.
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{contjpn}},
\code{\link{contmslm}},
\code{\link{getlang}},
\code{\link{othbest}},
\code{\link{othlang1}},
\code{\link{othlang2}},
\code{\link{parlang}},
\code{\link{spklang}},
\code{\link{uselang}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
