% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_253.R
\name{othlang1}
\alias{othlang1}
\title{What other languages does r speak}
\source{
General Social Survey https://gss.norc.org
}
\description{
othlang1
}
\details{
Question B. What other language(s) do you speak?  First response.
}
\section{Values}{

\itemize{
\item \code{1} english
\item \code{2} spanish
\item \code{3} polish
\item \code{4} french
\item \code{5} hebrew
\item \code{6} russian
\item \code{7} tagalog
\item \code{8} chinese
\item \code{9} urdu
\item \code{10} italian
\item \code{11} romanian
\item \code{12} german
\item \code{13} pashtu
\item \code{14} chamorro
\item \code{15} ukranian
\item \code{16} hungarian
\item \code{17} turkish
\item \code{18} khmer
\item \code{19} korean
\item \code{20} navajo
\item \code{21} malayam
\item \code{22} arabic
\item \code{23} farsi/persian
\item \code{24} serbian
\item \code{25} lithuanian
\item \code{27} thai
\item \code{28} gujarati
\item \code{30} dari
\item \code{31} slovak
\item \code{32} japanese
\item \code{33} portuguese
\item \code{34} vajun
\item \code{35} yiddish
\item \code{36} creole
\item \code{37} vietnamese
\item \code{38} norwegian
\item \code{39} malaysian
\item \code{40} latin
\item \code{41} czech
\item \code{42} hindi
\item \code{43} hakka
\item \code{44} swedish
\item \code{45} vesaya
\item \code{46} sign language
\item \code{47} tongan
\item \code{48} armenian
\item \code{49} ibo
\item \code{50} dutch
\item \code{51} cherokee
\item \code{52} samoan
\item \code{53} yoruba
\item \code{54} greek
\item \code{55} lakota
\item \code{56} african
\item \code{57} anishnabee
\item \code{58} ilokand
\item \code{59} assyrian
\item \code{60} danish
\item \code{61} laotian
\item \code{62} finnish
\item \code{63} mohawk
\item \code{64} amharic
\item \code{65} bengali
\item \code{66} oriya
\item \code{67} hawaiian
\item \code{68} albanian
\item \code{69} hmong
\item \code{70} indonesian
\item \code{71} swahili
\item \code{72} telugu
\item \code{73} kannada
\item \code{74} bulgarian
\item \code{75} guarani
\item \code{76} bosnian
\item \code{77} punjabi
\item \code{78} latvian
\item \code{79} basque
\item \code{80} bikol/bicolano
\item \code{81} croatian
\item \code{82} romani
\item \code{83} slovenian
\item \code{84} belgian
\item \code{85} mandingo
\item \code{86} jamaican
\item \code{87} somali
\item \code{88} other native american language
\item \code{89} tigrinya
\item \code{90} krio
\item \code{91} unspecified asian indian language
\item \code{92} cebuano
\item \code{93} macedonian
\item \code{94} patois
\item \code{95} pangasinan
\item \code{96} ilocano
\item \code{97} marathi
\item \code{98} kashmiri
\item \code{99} tamil
\item \code{100} akan
\item \code{101} ethiopian
\item \code{102} georgian
\item \code{103} yugoslavian
\item \code{104} irish
\item \code{105} catalan
\item \code{106} mixtec
\item \code{107} garifuna
\item \code{108} fang
\item \code{109} apache
\item \code{110} bassa
\item \code{111} efik
\item \code{112} twi
\item \code{113} maltese
\item \code{114} zuni
\item \code{115} ga
\item \code{116} mina
\item \code{117} surinamese
\item \code{118} yaqui
\item \code{119} quechua
\item \code{120} tok pisin
\item \code{121} pima
\item \code{122} lingala
\item \code{123} otomi
\item \code{124} chindali
\item \code{125} tibetan
\item \code{126} oromo
\item \code{127} nepali
\item \code{128} eskimo
\item \code{129} mam
\item \code{130} choctaw
\item \code{131} ashanti
\item \code{132} burmese
\item \code{133} hewa
\item \code{134} sango
\item \code{135} chitumbuka
\item \code{136} maratsi
\item \code{137} yemba
\item \code{138} k'iche'
\item \code{139} fante
\item \code{140} sinhalase
\item \code{141} icelandic
\item \code{142} maricopa
\item \code{143} quechan
\item \code{144} pohnpeian
\item \code{145} luganda
\item \code{146} shona
\item \code{147} urhobo
\item \code{148} colville tribe
\item \code{149} chickasau
\item \code{150} dakota
\item \code{151} hokkien
\item \code{152} wolof
\item \code{153} ewe
\item \code{154} igbo
\item \code{155} luhya
\item \code{156} black feet
\item \code{157} inupiaq
\item \code{158} paiute
\item \code{159} afrikaans
\item \code{160} kaien
\item \code{161} shawnee
\item \code{162} welsh
\item \code{163} chuj
\item \code{164} oremagna
\item \code{165} kaugel
\item \code{166} sanskrit
\item \code{167} hausa
\item \code{168} aramaic
\item \code{169} chuukese
\item \code{170} dagbani
\item \code{171} fijian
\item \code{172} kapampagng
\item \code{173} kinyarwanda
\item \code{174} doric
\item \code{175} sesotho
\item \code{176} sousou
\item \code{177} estonian
\item \code{178} ebon
\item \code{179} chibemba
\item \code{180} chewa
\item \code{181} gusii
\item \code{182} gun
\item \code{183} nuer
\item \code{184} oro
\item \code{185} kamba
\item \code{186} huichol
\item \code{187} papiamento
\item \code{188} hiligaynon
\item \code{189} mongolian
\item \code{190} bunong
\item \code{191} kikuyu
\item \code{192} konkani
\item \code{193} mende
\item \code{194} naija
\item \code{195} wendish
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2284/vshow}{GSS Data Explorer page for \code{othlang1}}.
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{contjpn}},
\code{\link{contmslm}},
\code{\link{getlang}},
\code{\link{othbest}},
\code{\link{othlang}},
\code{\link{othlang2}},
\code{\link{parlang}},
\code{\link{spklang}},
\code{\link{uselang}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
