% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_144.R
\name{othpay}
\alias{othpay}
\title{Get commissions, bonuses, tips from main job}
\source{
General Social Survey https://gss.norc.org
}
\description{
othpay
}
\details{
Question Do you receive any of the earnings from your main job in the form of commissions, bonuses, tips, or overtime?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1349/vshow}{GSS Data Explorer page for \code{othpay}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1991 \tab 605 \tab 377 \tab 523 \tab 12 \tab 1517 \cr
   Total \tab 605 \tab 377 \tab 523 \tab 12 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{fringe1}},
\code{\link{fringe2}},
\code{\link{fringe3}},
\code{\link{fringe4}},
\code{\link{fringe5}},
\code{\link{fringe6}},
\code{\link{raiseall}},
\code{\link{raisefav}},
\code{\link{raisehrd}}
}
\concept{Work Organizations}
\keyword{variable}
