% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_318.R
\name{othpet}
\alias{othpet}
\title{Other pet}
\source{
General Social Survey https://gss.norc.org
}
\description{
othpet
}
\details{
Question Current pet - Some other kind of pet
}
\section{Values}{

\itemize{
\item \code{0} does not have
\item \code{1} has
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7206/vshow}{GSS Data Explorer page for \code{othpet}}.
}

\keyword{variable}
