% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_580.R
\name{othrel1}
\alias{othrel1}
\title{What type of religious events or meetings}
\source{
General Social Survey https://gss.norc.org
}
\description{
othrel1
}
\details{
Question (Besides the regular, weekly worship service,) What type of religious events or meetings did you attend during the last seven days?
\enumerate{
\item First mention
}
}
\section{Values}{

\itemize{
\item \code{1} bible study,class
\item \code{2} choir, choral, concert, play
\item \code{3} discussion group
\item \code{4} lecture
\item \code{5} governing board, etc
\item \code{6} preaching, taught sunday school
\item \code{7} prayer group, breakfast, meeting
\item \code{8} counseling
\item \code{9} worship service
\item \code{10} other, religious
\item \code{11} other, non-religious
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/C/- \tab partial \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5321/vshow}{GSS Data Explorer page for \code{othrel1}}.
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{attreg}},
\code{\link{friday}},
\code{\link{mediarel}},
\code{\link{othrel}},
\code{\link{othrel2}},
\code{\link{saturday}},
\code{\link{thursday}},
\code{\link{tuesday}},
\code{\link{wednesdy}}

Other Religion: 
\code{\link{attreg}},
\code{\link{friday}},
\code{\link{mediarel}},
\code{\link{othrel}},
\code{\link{othrel2}},
\code{\link{saturday}},
\code{\link{thursday}},
\code{\link{tuesday}},
\code{\link{wednesdy}}
}
\concept{Religion}
\concept{Religious Attendance and Identity}
\keyword{variable}
