% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_580.R
\name{othrel2}
\alias{othrel2}
\title{What type of religious events or meetings}
\source{
General Social Survey https://gss.norc.org
}
\description{
othrel2
}
\details{
Question (Besides the regular, weekly worship service,) What type of religious events or meetings did you attend during the last seven days?
2. Second Mention
}
\section{Values}{

\itemize{
\item \verb{[1]} bible study,class
\item \verb{[2]} choir, choral, concert, play
\item \verb{[3]} discussion group
\item \verb{[4]} lecture
\item \verb{[5]} governing board, etc
\item \verb{[6]} preaching, taught sunday school
\item \verb{[7]} prayer group, breakfast, meeting
\item \verb{[8]} counseling
\item \verb{[9]} worship service
\item \verb{[10]} other, religious
\item \verb{[11]} other, non-religious
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   A/C/- \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5322/vshow}{GSS Data Explorer page for \code{othrel2}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab bible study,class \tab choir, choral, concert, play \tab discussion group \tab lecture \tab governing board, etc \tab preaching, taught sunday school \tab prayer group, breakfast, meeting \tab worship service \tab other, religious \tab Total \cr
   1996 \tab 2881 \tab 8 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 6 \tab 2904 \cr
   1998 \tab 2813 \tab 4 \tab 2 \tab 1 \tab 1 \tab - \tab 1 \tab 7 \tab - \tab 3 \tab 2832 \cr
   Total \tab 5694 \tab 12 \tab 3 \tab 2 \tab 2 \tab 1 \tab 2 \tab 10 \tab 1 \tab 9 \tab 5736 \cr
}
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{attreg}},
\code{\link{friday}},
\code{\link{mediarel}},
\code{\link{othrel}},
\code{\link{othrel1}},
\code{\link{saturday}},
\code{\link{thursday}},
\code{\link{tuesday}},
\code{\link{wednesdy}}
}
\concept{Religious Attendance and Identity}
\keyword{variable}
