% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_124.R
\name{othsch}
\alias{othsch}
\title{Children attend other school}
\source{
General Social Survey https://gss.norc.org
}
\description{
othsch
}
\details{
Question Do/did all of your children attend public school for all of their elementary and high school education?
A. Which types of school do/did your children attend?
6. Other
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1154/vshow}{GSS Data Explorer page for \code{othsch}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1998 \tab 2634 \tab 16 \tab 139 \tab 43 \tab 2832 \cr
   2000 \tab 2540 \tab 20 \tab 229 \tab 28 \tab 2817 \cr
   Total \tab 5174 \tab 36 \tab 368 \tab 71 \tab 5649 \cr
}
}

\seealso{
Other Religion: 
\code{\link{chrissch}},
\code{\link{denomsch}},
\code{\link{implives}},
\code{\link{othchrch}},
\code{\link{privsch}},
\code{\link{relsch}},
\code{\link{shopmove}},
\code{\link{shopnum}},
\code{\link{shoprel}}
}
\concept{Religion}
\keyword{variable}
