% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_112.R
\name{othself}
\alias{othself}
\title{Was this prob of narrow or wide concern}
\source{
General Social Survey https://gss.norc.org
}
\description{
othself
}
\details{
Question What about some representatives or governmental officials outside of the local community -- on the (county, IF LOCAL UNIT BELOW COUNTY LEVEL), state, or national level? Have you ever contacted or written to such a person on some need or problem?
A. Was this need or problem primarily of concern to you, your friends, and family or was it an issue of wider concern?
}
\section{Values}{

\itemize{
\item \verb{[1]} self frnds famly
\item \verb{[2]} both
\item \verb{[3]} wider concern
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1034/vshow}{GSS Data Explorer page for \code{othself}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab self frnds famly \tab both \tab wider concern \tab no answer \tab Total \cr
   1987 \tab 1318 \tab 129 \tab 48 \tab 312 \tab 12 \tab 1819 \cr
   Total \tab 1318 \tab 129 \tab 48 \tab 312 \tab 12 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{fedtrust}},
\code{\link{governor}},
\code{\link{loclobby}},
\code{\link{locself}},
\code{\link{locvote}},
\code{\link{othlobby}},
\code{\link{polrally}},
\code{\link{schlhead}},
\code{\link{usrep}}
}
\concept{Socio-political participation}
\keyword{variable}
