% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_525.R
\name{othssame}
\alias{othssame}
\title{No saving environment unless others do same}
\source{
General Social Survey https://gss.norc.org
}
\description{
othssame
}
\details{
Question How much do you agree or disagree with each of these statements?
B. There is no point in doing what I can for the environment unless others do the same.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010, 2021 \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4780/vshow}{GSS Data Explorer page for \code{othssame}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2000 \tab 1541 \tab 51 \tab 271 \tab 201 \tab 544 \tab 122 \tab 17 \tab 70 \tab - \tab 2817 \cr
   2010 \tab 614 \tab 66 \tab 335 \tab 121 \tab 688 \tab 189 \tab 23 \tab 8 \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 67 \tab 257 \tab 376 \tab 826 \tab 265 \tab 49 \tab 180 \tab 9 \tab 4032 \cr
   Total \tab 4158 \tab 184 \tab 863 \tab 698 \tab 2058 \tab 576 \tab 89 \tab 258 \tab 9 \tab 8893 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{amprogrn}},
\code{\link{bizpeop}},
\code{\link{genegen}},
\code{\link{govtbiz}},
\code{\link{grnexagg}},
\code{\link{grnintl}},
\code{\link{impgrn}},
\code{\link{peopgovt}},
\code{\link{popgrwth}}
}
\concept{ISSP Environment}
\keyword{variable}
