% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_239.R
\name{othyrs}
\alias{othyrs}
\title{Years worked full time at other parts of org}
\source{
General Social Survey https://gss.norc.org
}
\description{
othyrs
}
\details{
Question B. How many years in total (at those parts) have you worked?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2142/vshow}{GSS Data Explorer page for \code{othyrs}}.
}

\seealso{
Other Jobs: 
\code{\link{formltrn}},
\code{\link{genderwk}},
\code{\link{locyrs}},
\code{\link{orgsize}},
\code{\link{otjtrain}},
\code{\link{partorg}},
\code{\link{samework}},
\code{\link{siteyrs}},
\code{\link{wrkoth}}
}
\concept{Jobs}
\keyword{variable}
