% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_135.R
\name{outofbiz}
\alias{outofbiz}
\title{The organization went out of business}
\source{
General Social Survey https://gss.norc.org
}
\description{
outofbiz
}
\details{
Question In the last five years, did any of the following things happen to an organization that you worked for?
A. The organization went out of business or closed down?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1267/vshow}{GSS Data Explorer page for \code{outofbiz}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1991 \tab - \tab 149 \tab 1303 \tab 65 \tab 1517 \cr
   Total \tab 0 \tab 149 \tab 1303 \tab 65 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{genecanx}},
\code{\link{genecany}},
\code{\link{genegets}},
\code{\link{merged}},
\code{\link{profits1}},
\code{\link{profits2}},
\code{\link{reorg}},
\code{\link{unpower}},
\code{\link{unprog}}
}
\concept{Work Organizations}
\keyword{variable}
