% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_331.R
\name{outsider}
\alias{outsider}
\title{Getting mental health treatment make child an}
\source{
General Social Survey https://gss.norc.org
}
\description{
outsider
}
\details{
Question That was my last question about (John/Mary). Now, I have some questions about children, but not about
(John/Mary). Please tell me whether you Strongly Agree, Agree Somewhat, Disagree Somewhat or Strongly disagree with the following statements:
A. Getting mental health treatment for a child would make him/her an outsider at school.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree somewhat
\item \code{3} disagree somewhat
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2957/vshow}{GSS Data Explorer page for \code{outsider}}.
}

\seealso{
Other Mental Health II: 
\code{\link{comknows}},
\code{\link{failure}},
\code{\link{illphys}},
\code{\link{medkdneg}},
\code{\link{ovrmedkd}},
\code{\link{putsoff}},
\code{\link{sufadult}},
\code{\link{violpeop}},
\code{\link{violself}}
}
\concept{Mental Health II}
\keyword{variable}
