% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_609.R
\name{oversamp}
\alias{oversamp}
\title{Weights for black oversamples}
\source{
General Social Survey https://gss.norc.org
}
\description{
oversamp
}
\details{
Question Weight for Black oversamples.
}
\section{Values}{

\itemize{
\item \verb{[1]} not 1982/1987
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\seealso{
Other Sampling and Weights: 
\code{\link{cshutyp06}},
\code{\link{phase}},
\code{\link{res2006}},
\code{\link{res2008}},
\code{\link{res2010}},
\code{\link{res2012}},
\code{\link{spaneng}},
\code{\link{spanint}},
\code{\link{spanself}}
}
\concept{Sampling and Weights}
\keyword{variable}
