% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_183.R
\name{ovrjoyed}
\alias{ovrjoyed}
\title{Felt overjoyed about something?}
\source{
General Social Survey https://gss.norc.org
}
\description{
ovrjoyed
}
\details{
Question Now I'm going to read a list of different feelings that people sometimes have. After each one, I would like you to tell me on how many days you have felt this way during the past 7 days. On how many days in the past 7 days have you... (SPECIFY NUMBER OF DAYS)
(IF RESPONDENT ASKS, SAY THAT FEELING THIS WAY AT ANY TIME DURING A DAY MAKES THAT DAY "COUNT.")
J. Felt overjoyed about something?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1715/vshow}{GSS Data Explorer page for \code{ovrjoyed}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 584 \tab 264 \tab 185 \tab 145 \tab 73 \tab 56 \tab 30 \tab 107 \tab 12 \tab 4 \tab 2904 \cr
   Total \tab 1444 \tab 584 \tab 264 \tab 185 \tab 145 \tab 73 \tab 56 \tab 30 \tab 107 \tab 12 \tab 4 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{anxious}},
\code{\link{atease}},
\code{\link{contentd}},
\code{\link{excited}},
\code{\link{fearful}},
\code{\link{lonely}},
\code{\link{madat}},
\code{\link{restless}},
\code{\link{worried}}
}
\concept{Emotions}
\keyword{variable}
