% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_443.R
\name{ownautos}
\alias{ownautos}
\title{Govts role in automobile industry}
\source{
General Social Survey https://gss.norc.org
}
\description{
ownautos
}
\details{
Question What do you think the government''s role in each of these industries should be?
E. The automobile industry.
}
\section{Values}{

\itemize{
\item \code{1} own it
\item \code{2} control prices
\item \code{3} leave alone
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4075/vshow}{GSS Data Explorer page for \code{ownautos}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab own it \tab control prices \tab leave alone \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 14 \tab 226 \tab 373 \tab 46 \tab 18 \tab 1534 \cr
   Total \tab 857 \tab 14 \tab 226 \tab 373 \tab 46 \tab 18 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{buspow}},
\code{\link{govtpow}},
\code{\link{jobsall}},
\code{\link{laborpow}},
\code{\link{ownbanks}},
\code{\link{ownmass}},
\code{\link{ownpower}},
\code{\link{ownsteel}},
\code{\link{pricecon}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
