% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_417.R
\name{ownbiz}
\alias{ownbiz}
\title{R owns business}
\source{
General Social Survey https://gss.norc.org
}
\description{
ownbiz
}
\details{
Question Are you, alone or with others, currently the owner of a business you help manage, including self-employment or selling any goods or services to others?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3815/vshow}{GSS Data Explorer page for \code{ownbiz}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2008 \tab - \tab 250 \tab 1771 \tab 2 \tab 2023 \cr
   Total \tab 0 \tab 250 \tab 1771 \tab 2 \tab 2023 \cr
}
}

\seealso{
Other Self Employment: 
\code{\link{numown}},
\code{\link{startbiz}},
\code{\link{yearbiz}}
}
\concept{Self Employment}
\keyword{variable}
