% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_679.R
\name{owndecns}
\alias{owndecns}
\title{Best if every nation had its own country}
\source{
General Social Survey https://gss.norc.org
}
\description{
owndecns
}
\details{
Question Thinking in general about nations and countries all over the world, to what extent do you agree or disagree with the following statement?… It would be best if every nation had its own country where it could make its own decisions.
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} disagree strongly
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8066/vshow}{GSS Data Explorer page for \code{owndecns}}.
}

\concept{ISSP Citizenship}
\keyword{variable}
