% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_340.R
\name{owndoing}
\alias{owndoing}
\title{Degree of influence over what happens to me}
\source{
General Social Survey https://gss.norc.org
}
\description{
owndoing
}
\details{
Question Please indicate which of the following statements comes closer to describing you by circling the appropriate letter:
}
\section{Values}{

\itemize{
\item \code{1} i have little influence over things that happens to me
\item \code{2} what happens to me is largely my own doing
\item \code{9} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3045/vshow}{GSS Data Explorer page for \code{owndoing}}.
}

\seealso{
Other Genes and the Environment: 
\code{\link{athletic}},
\code{\link{depndabl}},
\code{\link{geneexps}},
\code{\link{intrwght}},
\code{\link{kindpers}},
\code{\link{moregood}},
\code{\link{sadblue}},
\code{\link{selfish}}

Other Satisfaction: 
\code{\link{athletic}},
\code{\link{depndabl}},
\code{\link{kindpers}},
\code{\link{moregood}},
\code{\link{sadblue}},
\code{\link{selfish}}
}
\concept{Genes and the Environment}
\concept{Satisfaction}
\keyword{variable}
