% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_174.R
\name{owneth}
\alias{owneth}
\title{Do ethic politicians serve ethnic groups better}
\source{
General Social Survey https://gss.norc.org
}
\description{
owneth
}
\details{
Question Do you feel that people are best represented in politics by leaders from their own racial or ethnic background, or doesn't the leader's background make very much difference?
}
\section{Values}{

\itemize{
\item \code{1} leaders from own background
\item \code{2} leaders background no matter
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1626/vshow}{GSS Data Explorer page for \code{owneth}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab leaders from own background \tab leaders background no matter \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 497 \tab 914 \tab 51 \tab 12 \tab 2992 \cr
   Total \tab 1518 \tab 497 \tab 914 \tab 51 \tab 12 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{congeth}},
\code{\link{discaffy}},
\code{\link{ethhist}},
\code{\link{jobaff}},
\code{\link{schleth}},
\code{\link{teacheth}},
\code{\link{whoteach}}
}
\concept{Multiculturalism}
\keyword{variable}
