% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_512.R
\name{ownfate}
\alias{ownfate}
\title{We each make our own fate}
\source{
General Social Survey https://gss.norc.org
}
\description{
ownfate
}
\details{
Question Do you agree or disagree with the following. . .
G. We each make our own fate.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4663/vshow}{GSS Data Explorer page for \code{ownfate}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 266 \tab 517 \tab 254 \tab 145 \tab 59 \tab 50 \tab 68 \tab 1517 \cr
   1998 \tab 1548 \tab 200 \tab 510 \tab 237 \tab 149 \tab 79 \tab 44 \tab 65 \tab 2832 \cr
   Total \tab 1706 \tab 466 \tab 1027 \tab 491 \tab 294 \tab 138 \tab 94 \tab 133 \tab 4349 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{denkid}},
\code{\link{egomeans}},
\code{\link{madenkid}},
\code{\link{marelkid}},
\code{\link{padenkid}},
\code{\link{parelkid}},
\code{\link{relexper}},
\code{\link{religkid}},
\code{\link{religsp}}
}
\concept{ISSP Religion}
\keyword{variable}
