% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_418.R
\name{owninc}
\alias{owninc}
\title{Amt of household income coming from this business}
\source{
General Social Survey https://gss.norc.org
}
\description{
owninc
}
\details{
Question In which of these groups did your earnings from this business fall in contributing towards your household income For last year?
IF R HAS MORE THAN ONE BUSINESS/FARM/PARTNERSHIP, ASK ABOUT THE PRIMARY BUSINESS THEY ARE INVOLVED IN.
}
\section{Values}{

\itemize{
\item \code{1} ( $15,000
\item \code{2} $15,001-$25,000
\item \code{3} $25,001-$30,000
\item \code{4} $30,001-$35,000
\item \code{5} $35,001-$50,000
\item \code{6} $50,001-$60,000
\item \code{7} $60,001-$75,000
\item \code{8} $75,001-$100,000
\item \code{9} ) $100,000
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3821/vshow}{GSS Data Explorer page for \code{owninc}}.
}

\seealso{
Other Self Employment: 
\code{\link{bizgross}},
\code{\link{bizshare}},
\code{\link{numemps5}},
\code{\link{paidemps}},
\code{\link{whybiz}}
}
\concept{Self Employment}
\keyword{variable}
