% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_443.R
\name{ownpower}
\alias{ownpower}
\title{Govts role in electric power}
\source{
General Social Survey https://gss.norc.org
}
\description{
ownpower
}
\details{
Question What do you think the government''s role in each of these industries should be?
A. Electric power.
}
\section{Values}{

\itemize{
\item \verb{[1]} own it
\item \verb{[2]} control prices
\item \verb{[3]} leave alone
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4071/vshow}{GSS Data Explorer page for \code{ownpower}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab own it \tab control prices \tab leave alone \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 37 \tab 394 \tab 199 \tab 34 \tab 13 \tab 1534 \cr
   1990 \tab 155 \tab 63 \tab 618 \tab 366 \tab 122 \tab 48 \tab 1372 \cr
   Total \tab 1012 \tab 100 \tab 1012 \tab 565 \tab 156 \tab 61 \tab 2906 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{buspow}},
\code{\link{govtpow}},
\code{\link{jobsall}},
\code{\link{laborpow}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{ownmass}},
\code{\link{ownsteel}},
\code{\link{pricecon}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
