% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_443.R
\name{ownpower}
\alias{ownpower}
\title{Govts role in electric power}
\source{
General Social Survey https://gss.norc.org
}
\description{
ownpower
}
\details{
Question What do you think the government''s role in each of these industries should be?
A. Electric power.
}
\section{Values}{

\itemize{
\item \code{1} own it
\item \code{2} control prices
\item \code{3} leave alone
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4071/vshow}{GSS Data Explorer page for \code{ownpower}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{buspow}},
\code{\link{govtpow}},
\code{\link{jobsall}},
\code{\link{laborpow}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{ownmass}},
\code{\link{ownsteel}},
\code{\link{pricecon}}

Other Business: 
\code{\link{buspow}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{ownmass}},
\code{\link{ownsteel}}

Other Economy: 
\code{\link{jobsall}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{ownmass}},
\code{\link{ownsteel}},
\code{\link{pricecon}}

Other Federal Government: 
\code{\link{govtpow}},
\code{\link{jobsall}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{ownmass}},
\code{\link{ownsteel}},
\code{\link{pricecon}}
}
\concept{Business}
\concept{Economy}
\concept{Federal Government}
\concept{ISSP Role of Gov't}
\keyword{variable}
