% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_443.R
\name{ownsteel}
\alias{ownsteel}
\title{Govts role in the steel industry}
\source{
General Social Survey https://gss.norc.org
}
\description{
ownsteel
}
\details{
Question What do you think the government''s role in each of these industries should be?
C. The steel industry.
}
\section{Values}{

\itemize{
\item \verb{[1]} own it
\item \verb{[2]} control prices
\item \verb{[3]} leave alone
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4073/vshow}{GSS Data Explorer page for \code{ownsteel}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab own it \tab control prices \tab leave alone \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 15 \tab 216 \tab 371 \tab 53 \tab 22 \tab 1534 \cr
   1990 \tab 155 \tab 21 \tab 364 \tab 623 \tab 152 \tab 57 \tab 1372 \cr
   Total \tab 1012 \tab 36 \tab 580 \tab 994 \tab 205 \tab 79 \tab 2906 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{buspow}},
\code{\link{govtpow}},
\code{\link{jobsall}},
\code{\link{laborpow}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{ownmass}},
\code{\link{ownpower}},
\code{\link{pricecon}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
