% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_320.R
\name{ownstock}
\alias{ownstock}
\title{R has stock in r's company}
\source{
General Social Survey https://gss.norc.org
}
\description{
ownstock
}
\details{
Question Do you own any shares of stock in the company where you now work, either directly or through some type of retirement or stock plan?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no, i work for a company that does not have stock
\item \code{3} no, my company has stock but i do not own any
\item \code{4} no, i do not own and do not know if my company has stock
\item \code{5} i do not know if i own stock in my company
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2852/vshow}{GSS Data Explorer page for \code{ownstock}}.
}

\seealso{
Other Employee Compensation: 
\code{\link{stockops}},
\code{\link{stockval}},
\code{\link{workfor}},
\code{\link{workfor1}}

Other Job: 
\code{\link{stockops}},
\code{\link{stockval}}
}
\concept{Employee Compensation}
\concept{Job}
\keyword{variable}
