% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_115.R
\name{paden}
\alias{paden}
\title{What was father's religious denomination}
\source{
General Social Survey https://gss.norc.org
}
\description{
paden
}
\details{
Question A. What specific denomination is that, if any? (PROBE FOR EXACT DENOMINATION)
}
\section{Values}{

\itemize{
\item \code{10} am baptist asso
\item \code{11} am bapt ch in usa
\item \code{12} nat bapt conv of am
\item \code{13} nat bapt conv usa
\item \code{14} southern baptist
\item \code{15} other baptists
\item \code{18} baptist-dk which
\item \code{20} afr meth episcopal
\item \code{21} afr meth ep zion
\item \code{22} united methodist
\item \code{23} other methodist
\item \code{28} methodist-dk which
\item \code{30} am lutheran
\item \code{31} luth ch in america
\item \code{32} lutheran-mo synod
\item \code{33} wi evan luth synod
\item \code{34} other lutheran
\item \code{35} evangelical luth
\item \code{38} lutheran-dk which
\item \code{40} presbyterian c in us
\item \code{41} united pres ch in usa
\item \code{42} other presbyterian
\item \code{48} presbyterian-dk wh
\item \code{50} episcopal
\item \code{60} other
\item \code{70} no denomination
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1064/vshow}{GSS Data Explorer page for \code{paden}}.

Counts by year:\tabular{lllllllllllllllllllllllllllll}{
   year \tab iap \tab am baptist asso \tab am bapt ch in usa \tab nat bapt conv of am \tab nat bapt conv usa \tab southern baptist \tab other baptists \tab baptist-dk which \tab afr meth episcopal \tab afr meth ep zion \tab united methodist \tab other methodist \tab methodist-dk which \tab am lutheran \tab luth ch in america \tab lutheran-mo synod \tab wi evan luth synod \tab other lutheran \tab lutheran-dk which \tab presbyterian c in us \tab united pres ch in usa \tab other presbyterian \tab presbyterian-dk wh \tab episcopal \tab other \tab no denomination \tab no answer \tab Total \cr
   1988 \tab 772 \tab 19 \tab 5 \tab 1 \tab 6 \tab 128 \tab 30 \tab 66 \tab 4 \tab 3 \tab 88 \tab 8 \tab 25 \tab 21 \tab 11 \tab 12 \tab 1 \tab 4 \tab 27 \tab 9 \tab 15 \tab 6 \tab 28 \tab 24 \tab 122 \tab 37 \tab 9 \tab 1481 \cr
   Total \tab 772 \tab 19 \tab 5 \tab 1 \tab 6 \tab 128 \tab 30 \tab 66 \tab 4 \tab 3 \tab 88 \tab 8 \tab 25 \tab 21 \tab 11 \tab 12 \tab 1 \tab 4 \tab 27 \tab 9 \tab 15 \tab 6 \tab 28 \tab 24 \tab 122 \tab 37 \tab 9 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{churhsch}},
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{maoth}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{paoth}},
\code{\link{parelig}},
\code{\link{sunsch16}}
}
\concept{Religion}
\keyword{variable}
