% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_303.R
\name{paidhow}
\alias{paidhow}
\title{How are you paid}
\source{
General Social Survey https://gss.norc.org
}
\description{
paidhow
}
\details{
Question In this section, we ask about your work hours and how you are paid.  When referring to jobs, we are referring to jobs for which you work for pay, not volunteer work, unpaid housework, or other unpaid activities. When we refer to work hours, we ask you to include all hours performed for a paid job, including any extra hours, overtime, work you did at home for your paid job, and hours that may not have been directly billable or compensated.
Are you salaried, paid by the hour, or paid some other way? If you have more than one job, think of the job in which you spend the most hours working each week.
}
\section{Values}{

\itemize{
\item \code{1} salary
\item \code{2} by the hour
\item \code{3} earnings from business
\item \code{4} piecework
\item \code{5} by the job/project
\item \code{6} wage + commission
\item \code{7} by hour + tips
\item \code{8} mileage
\item \code{9} salary & hours combination
\item \code{10} by day/week
\item \code{11} by route
\item \code{12} by client
\item \code{13} by contract
\item \code{20} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6291/vshow}{GSS Data Explorer page for \code{paidhow}}.
}

\seealso{
Other Flexible Work: 
\code{\link{mosthrs}},
\code{\link{usualhrs}}
}
\concept{Flexible Work}
\keyword{variable}
