% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_489.R
\name{paidlv}
\alias{paidlv}
\title{Paid leave for childcare}
\source{
General Social Survey https://gss.norc.org
}
\description{
paidlv
}
\details{
Question Consider a couple who both work full-time and now have a new born child. One of them stops working for some time to care for their child. Do you think there should be paid leave available and, if so, for how long?
}
\section{Values}{

\itemize{
\item \code{0} no, there should be no paid leave
\item \code{1} yes, enter months
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B \tab partial \cr
   2022 \tab A/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4463/vshow}{GSS Data Explorer page for \code{paidlv}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{carecost}},
\code{\link{careprov}},
\code{\link{eldersup}},
\code{\link{eldhelp}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{kidsocst}},
\code{\link{paidlvdv}},
\code{\link{paidlvpy}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
