% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_489.R
\name{paidlvdv}
\alias{paidlvdv}
\title{Mother or father paid leave}
\source{
General Social Survey https://gss.norc.org
}
\description{
paidlvdv
}
\details{
Question Still thinking about the same couple, if both are in a similar work situation and are eligible for paid leave, how should this paid leave period be divided between the mother and the father?
}
\section{Values}{

\itemize{
\item \code{1} the mother should take the entire paid leave period and the father should not take any paid leave
\item \code{2} the mother should take most of the paid leave period and the father should take some of it
\item \code{3} the mother and the father should each take half of the paid leave period
\item \code{4} the father should take most of the paid leave period and the mother should take some of it
\item \code{5} the father should take the entire paid leave period and the mother should not take any paid leave
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B \tab partial \cr
   2022 \tab A/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4466/vshow}{GSS Data Explorer page for \code{paidlvdv}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{carecost}},
\code{\link{careprov}},
\code{\link{eldersup}},
\code{\link{eldhelp}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{kidsocst}},
\code{\link{paidlv}},
\code{\link{paidlvpy}}

Other Children: 
\code{\link{eldersup}},
\code{\link{kidsocst}},
\code{\link{paidlv}},
\code{\link{paidlvpy}}
}
\concept{Children}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
