% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_489.R
\name{paidlvdv}
\alias{paidlvdv}
\title{Mother or father paid leave}
\source{
General Social Survey https://gss.norc.org
}
\description{
paidlvdv
}
\details{
Question Still thinking about the same couple, if both are in a similar work situation and are eligible for paid leave, how should this paid leave period be divided between the mother and the father?
}
\section{Values}{

\itemize{
\item \code{1} the mother should take the entire paid leave period and the father should not take any paid leave
\item \code{2} the mother should take most of the paid leave period and the father should take some of it
\item \code{3} the mother and the father should each take half of the paid leave period
\item \code{4} the father should take most of the paid leave period and the mother should take some of it
\item \code{5} the father should take the entire paid leave period and the mother should not take any paid leave
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B \tab partial \cr
   2022 \tab A/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4466/vshow}{GSS Data Explorer page for \code{paidlvdv}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{abchoose}},
\code{\link{buythngs}},
\code{\link{carecost}},
\code{\link{careprov}},
\code{\link{caresick}},
\code{\link{caresik1}},
\code{\link{chldcare}},
\code{\link{clean1}},
\code{\link{cohabfst}},
\code{\link{cohabit}},
\code{\link{cohabok}},
\code{\link{cooking1}},
\code{\link{daycare1}},
\code{\link{daycare2}},
\code{\link{daycare3}},
\code{\link{daycare4}},
\code{\link{daycare5}},
\code{\link{deckids}},
\code{\link{dinner}},
\code{\link{divbest}},
\code{\link{divhubby}},
\code{\link{divifkd1}},
\code{\link{divifkid}},
\code{\link{divkids}},
\code{\link{divnokd1}},
\code{\link{divnokid}},
\code{\link{divnow}},
\code{\link{divwife}},
\code{\link{earnshh}},
\code{\link{earnsmor}},
\code{\link{eldcost}},
\code{\link{eldersup}},
\code{\link{eldhelp}},
\code{\link{evdiv}},
\code{\link{fambudgt}},
\code{\link{famsuffr}},
\code{\link{famvswk1}},
\code{\link{famvswrk}},
\code{\link{famwkbst}},
\code{\link{famwklst}},
\code{\link{fejobind}},
\code{\link{femarry}},
\code{\link{fewknokd}},
\code{\link{fewrksup}},
\code{\link{fourkids}},
\code{\link{hapifwrk}},
\code{\link{happy7}},
\code{\link{hhclean1}},
\code{\link{hhwkdis}},
\code{\link{hhwkfair}},
\code{\link{homekid}},
\code{\link{housewrk}},
\code{\link{hubbywk1}},
\code{\link{hubbywrk}},
\code{\link{ilikejob}},
\code{\link{jobvsfa1}},
\code{\link{jobvsfam}},
\code{\link{kidempty}},
\code{\link{kidfin}},
\code{\link{kidfinbu}},
\code{\link{kidjob}},
\code{\link{kidjoy}},
\code{\link{kidless}},
\code{\link{kidnofre}},
\code{\link{kidsocst}},
\code{\link{kidsuffr}},
\code{\link{kidtrble}},
\code{\link{laundry}},
\code{\link{laundry1}},
\code{\link{livnowed}},
\code{\link{mapaid}},
\code{\link{mardiv}},
\code{\link{marfin}},
\code{\link{marfree}},
\code{\link{marhappy}},
\code{\link{markids}},
\code{\link{marlegit}},
\code{\link{marmakid}},
\code{\link{marnomar}},
\code{\link{marpakid}},
\code{\link{mawork14}},
\code{\link{mawrkwrm}},
\code{\link{mehhwork}},
\code{\link{mekdcare}},
\code{\link{memarry}},
\code{\link{meovrwrk}},
\code{\link{mrmom}},
\code{\link{nokids}},
\code{\link{numkids}},
\code{\link{onekid}},
\code{\link{paidlv}},
\code{\link{paidlvpy}},
\code{\link{repairs}},
\code{\link{repairs1}},
\code{\link{rfamlook}},
\code{\link{rhhwork}},
\code{\link{rwrkbaby}},
\code{\link{rwrkgrwn}},
\code{\link{rwrknokd}},
\code{\link{rwrksch}},
\code{\link{satfam7}},
\code{\link{satjob7}},
\code{\link{sexhar}},
\code{\link{shop1}},
\code{\link{shopfood}},
\code{\link{singlpar}},
\code{\link{spborn}},
\code{\link{spevdiv}},
\code{\link{spfalook}},
\code{\link{sphhwork}},
\code{\link{splive}},
\code{\link{spwkbaby}},
\code{\link{spwkgrwn}},
\code{\link{spwknokd}},
\code{\link{spwrksch}},
\code{\link{ssfchild}},
\code{\link{ssmchild}},
\code{\link{strsshme}},
\code{\link{strsswrk}},
\code{\link{teensex1}},
\code{\link{threkids}},
\code{\link{timehome}},
\code{\link{timework}},
\code{\link{tiredhm1}},
\code{\link{tiredhme}},
\code{\link{tiredwk1}},
\code{\link{tiredwrk}},
\code{\link{twoincs}},
\code{\link{twoincs1}},
\code{\link{twokids}},
\code{\link{weekend}},
\code{\link{wkkidscl}},
\code{\link{wkkidscs}},
\code{\link{wkndact}},
\code{\link{wkyngscl}},
\code{\link{wkyngscs}},
\code{\link{wrkbaby}},
\code{\link{wrkgrown}},
\code{\link{wrknokid}},
\code{\link{wrksch}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
