% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_296.R
\name{pain}
\alias{pain}
\title{Pain interfere with work past 4 wks}
\source{
General Social Survey https://gss.norc.org
}
\description{
pain
}
\details{
Question During the past 4 weeks, how much did pain interfere with your normal work (including both work outside the home and housework)?
}
\section{Values}{

\itemize{
\item \code{1} not at all
\item \code{2} a little bit
\item \code{3} moderately
\item \code{4} quite a bit
\item \code{5} extremely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2683/vshow}{GSS Data Explorer page for \code{pain}}.
}

\seealso{
Other Health Status: 
\code{\link{downblue}},
\code{\link{energy}},
\code{\link{peaceful}},
\code{\link{socacts}},
\code{\link{treat1}},
\code{\link{treat2}},
\code{\link{treat3}},
\code{\link{treat4}},
\code{\link{treat5}}
}
\concept{Health Status}
\keyword{variable}
