% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_311.R
\name{painarms}
\alias{painarms}
\title{R had pain in the arms in the past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
painarms
}
\details{
Question B. In the past 12 months, have you had pain in the hands, wrists, arms, or shoulders every day for a week or more?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2832/vshow}{GSS Data Explorer page for \code{painarms}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 504 \tab 1271 \tab 6 \tab 15 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 473 \tab 1249 \tab 1 \tab 11 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 318 \tab 842 \tab 1 \tab 26 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 305 \tab 939 \tab 1 \tab 5 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 406 \tab 1008 \tab - \tab 34 \tab - \tab 2348 \cr
   2022 \tab 1559 \tab 580 \tab 1364 \tab 5 \tab 16 \tab 20 \tab 3544 \cr
   Total \tab 8349 \tab 2586 \tab 6673 \tab 14 \tab 107 \tab 20 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{backpain}},
\code{\link{hlthdays}},
\code{\link{hurtatwk}},
\code{\link{mntlhlth}},
\code{\link{phyeffrt}},
\code{\link{physhlth}},
\code{\link{spvtrfair}},
\code{\link{strredpg}},
\code{\link{usedup}}
}
\concept{Quality of Working Life}
\keyword{variable}
