% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_07.R
\name{paind16}
\alias{paind16}
\title{Father's industry code (1970)}
\source{
General Social Survey https://gss.norc.org
}
\description{
paind16
}
\details{
Question A. What kind of work did your father (FATHER SUBSTITUTE) normally do while you were growing up? That is, what was his job called?
B. What did he actually do in that job? Tell me, what were some of his main duties?
C. What kind of place did he work for?
D. What did they (make/do)?
E. Was he self-employed or did he work for someone else?
}
\section{Values}{

\itemize{
\item \verb{[17]} agricultural production
\item \verb{[18]} agricultural services, except horticultural
\item \verb{[19]} horticultural services
\item \verb{[27]} forestry
\item \verb{[28]} fisheries
\item \verb{[47]} metal mining
\item \verb{[48]} coal mining
\item \verb{[49]} crude petroleum and natural gas extractions
\item \verb{[57]} nonmetallic mining and quarrying, except fuel
\item \verb{[67]} general building contractors
\item \verb{[68]} general contractors, except building
\item \verb{[69]} special trade contractors
\item \verb{[77]} not specified construction
\item \verb{[107]} logging
\item \verb{[108]} sawmills, planing mills, and  mill work
\item \verb{[109]} miscellaneous wood products
\item \verb{[118]} furniture fixtures
\item \verb{[119]} glass and glass products
\item \verb{[127]} cement concrete, gypsum, and  plaster products
\item \verb{[128]} structural clay products
\item \verb{[137]} pottery and related products
\item \verb{[138]} miscellaneous nonmetallic mineral and stone products
\item \verb{[139]} blast furnaces, steel works, rolling and finishing mills
\item \verb{[147]} other primary iron and steel industries
\item \verb{[148]} primary aluminum industries
\item \verb{[149]} other primary nonferrous industries
\item \verb{[157]} cutlery, hand tools, and other hardware
\item \verb{[158]} fabricated structural metal products
\item \verb{[159]} screw machine products
\item \verb{[167]} metal stamping
\item \verb{[168]} miscellaneous fabricated metal products
\item \verb{[169]} not specified metal industries
\item \verb{[177]} engines and turbines
\item \verb{[178]} farm machinery and equipment
\item \verb{[179]} construction and material handling machines
\item \verb{[187]} metalworking machinery
\item \verb{[188]} office and accounting machines
\item \verb{[189]} electronic computing equipment
\item \verb{[197]} machinery, except electrical, n.e.c.
\item \verb{[198]} not specified machinery
\item \verb{[199]} household appliances
\item \verb{[207]} radio, t.v., and communication equipment
\item \verb{[208]} electrical machinery, equipment, and supplies, n.e.c.
\item \verb{[209]} not specified electrical machinery, equipment, and supplies
\item \verb{[219]} motor vehicles and motor vehicle equipment
\item \verb{[227]} aircraft and parts
\item \verb{[228]} ship and boat building and repairing
\item \verb{[229]} railroad locomotives and equipment
\item \verb{[237]} mobile dwellings and campers
\item \verb{[238]} cycles and miscellaneous transportation equipment
\item \verb{[239]} scientific and controlling instruments
\item \verb{[247]} optical and health services supplies
\item \verb{[248]} photographic equipment and supplies
\item \verb{[249]} watches, clocks, and clockwork-operated devices
\item \verb{[257]} not specified professional equipment
\item \verb{[258]} ordinance
\item \verb{[259]} miscellaneous manufacturing industries
\item \verb{[268]} meat products
\item \verb{[269]} dairy products
\item \verb{[278]} canning and preserving fruits, vegetables, and sea foods
\item \verb{[279]} grain-mill products
\item \verb{[287]} bakery products
\item \verb{[288]} confectionery and related products
\item \verb{[289]} beverage industries
\item \verb{[297]} miscellaneous food preparation and kindred products
\item \verb{[298]} not specified food industries
\item \verb{[299]} tobacco manufactures
\item \verb{[307]} knitting mills
\item \verb{[308]} dyeing and finishing textiles, except wool and knit goods
\item \verb{[309]} floor coverings, except hard surface
\item \verb{[317]} yarn, thread, and fabric mills
\item \verb{[318]} miscellaneous textile mill products
\item \verb{[319]} apparel and accessories
\item \verb{[327]} miscellaneous fabricated textile products
\item \verb{[328]} pulp, paper and paperboard mills
\item \verb{[329]} miscellaneous paper and pulp products
\item \verb{[337]} paperboard containers and boxes
\item \verb{[338]} newspaper publishing and printing
\item \verb{[339]} printing, publishing, and allied industries, except newspapers
\item \verb{[347]} industrial chemicals
\item \verb{[348]} plastics, synthetics and resins, except fibers
\item \verb{[349]} synthetic fibers
\item \verb{[357]} drugs and medicines
\item \verb{[358]} soaps and cosmetics
\item \verb{[359]} paints, varnishes, and related products
\item \verb{[367]} agricultural chemicals
\item \verb{[368]} miscellaneous chemicals
\item \verb{[369]} not specified chemicals and allied products
\item \verb{[377]} petroleum refining
\item \verb{[378]} miscellaneous petroleum and coal products
\item \verb{[379]} rubber products
\item \verb{[387]} miscellaneous plastic products
\item \verb{[388]} tanned, curried, and finished leather
\item \verb{[389]} footwear, except rubber
\item \verb{[397]} leather products, except footwear
\item \verb{[398]} not specified manufacturing industries
\item \verb{[407]} railroads and railway express
\item \verb{[408]} street railways and bus lines
\item \verb{[409]} taxicab service
\item \verb{[417]} trucking service
\item \verb{[418]} warehousing and storage
\item \verb{[419]} water transportation
\item \verb{[427]} air transportation
\item \verb{[428]} pipe lines, except natural gas
\item \verb{[429]} services incidental to transportation
\item \verb{[447]} radio broadcasting and television
\item \verb{[448]} telephone (wire and radio)
\item \verb{[449]} telegraph and miscellaneous communication services
\item \verb{[467]} electric light and power
\item \verb{[468]} electric-gas utilities
\item \verb{[469]} gas and steam supply systems
\item \verb{[477]} water supply
\item \verb{[478]} sanitary services
\item \verb{[479]} other and not specified utilities
\item \verb{[507]} motor vehicles and equipment
\item \verb{[508]} drugs, chemicals, and allied products
\item \verb{[509]} dry goods and apparel
\item \verb{[527]} food and related products
\item \verb{[528]} farm products-raw materials
\item \verb{[529]} electrical goods
\item \verb{[537]} hardware, plumbing, and heating supplies
\item \verb{[538]} not specified electrical and hardware products
\item \verb{[539]} machinery equipment and supplies
\item \verb{[557]} metals and minerals, n.e.c.
\item \verb{[558]} petroleum products
\item \verb{[559]} scrap and waste materials
\item \verb{[567]} alcoholic beverages
\item \verb{[568]} paper and its products
\item \verb{[569]} lumber and construction materials
\item \verb{[587]} wholesalers, n.e.c.
\item \verb{[588]} not specified wholesale trade
\item \verb{[607]} lumber and building material retailing
\item \verb{[608]} hardware and farm equipment stores
\item \verb{[609]} department and mail order establishments
\item \verb{[617]} limited price variety stores
\item \verb{[618]} vending machine operators
\item \verb{[619]} direct selling establishments
\item \verb{[627]} misc. general merchandise stores
\item \verb{[628]} grocery stores
\item \verb{[629]} dairy product stores
\item \verb{[637]} retail bakeries
\item \verb{[638]} food stores, n.e.c.
\item \verb{[639]} motor vehicle dealers
\item \verb{[647]} tire, battery, and accessory dealers
\item \verb{[648]} gasoline service stations
\item \verb{[649]} miscellaneous vehicle dealers
\item \verb{[657]} apparel and accessories stores, except shoe stores
\item \verb{[658]} shoe stores
\item \verb{[667]} furniture and home furnishing stores
\item \verb{[668]} household appliances, tv, and radio stores
\item \verb{[669]} eating and drinking places
\item \verb{[677]} drug stores
\item \verb{[678]} liquor stores
\item \verb{[679]} farm and garden supply stores
\item \verb{[687]} jewelry stores
\item \verb{[688]} fuel and ice dealers
\item \verb{[689]} retail florists
\item \verb{[697]} miscellaneous retail stores
\item \verb{[698]} not specified retail trade
\item \verb{[707]} banking
\item \verb{[708]} credit agencies
\item \verb{[709]} security, commodity brokerage, and investment companies
\item \verb{[717]} insurance
\item \verb{[718]} real estate, incl. real estate-insurance-law offices
\item \verb{[727]} advertising
\item \verb{[728]} services to dwellings and other buildings
\item \verb{[729]} commercial research, development, and testing labs
\item \verb{[737]} employment and temporary help agencies
\item \verb{[738]} business management and consulting services
\item \verb{[739]} computer programming services
\item \verb{[747]} detective and protective services
\item \verb{[748]} business services, n.e.c.
\item \verb{[749]} automobile services, except repair
\item \verb{[757]} automobile repair and related services
\item \verb{[758]} electrical repair shops
\item \verb{[759]} miscellaneous repair services
\item \verb{[769]} private households
\item \verb{[777]} hotels and motels
\item \verb{[778]} lodging places, except hotels and motels
\item \verb{[779]} laundering, cleaning, and other garment services
\item \verb{[787]} beauty shops
\item \verb{[788]} barber shops
\item \verb{[789]} shoe repair shops
\item \verb{[797]} dressmaking shops
\item \verb{[798]} miscellaneous personal services
\item \verb{[807]} theaters and motion pictures
\item \verb{[808]} bowling alleys, billiard and pool parlors
\item \verb{[809]} miscellaneous entertainment and recreation services
\item \verb{[828]} offices of physicians
\item \verb{[829]} offices of dentists
\item \verb{[837]} offices of chiropractors
\item \verb{[838]} hospitals
\item \verb{[839]} convalescent institutions
\item \verb{[847]} offices of health practitioners,n.e.c.
\item \verb{[848]} health services, n.e.c.
\item \verb{[849]} legal services
\item \verb{[857]} elementary and secondary schools
\item \verb{[858]} colleges and universities
\item \verb{[859]} libraries
\item \verb{[867]} educational services, n.e.c.
\item \verb{[868]} not specified educational services
\item \verb{[869]} museums, art galleries, and zoos
\item \verb{[877]} religious organizations
\item \verb{[878]} welfare services
\item \verb{[879]} residential welfare facilities
\item \verb{[887]} nonprofit membership organizations
\item \verb{[888]} engineering and architectural services
\item \verb{[889]} accounting, auditing, and bookkeeping services
\item \verb{[897]} miscellaneous professional and related services
\item \verb{[907]} postal service
\item \verb{[917]} federal public administration
\item \verb{[927]} state public administration
\item \verb{[937]} local public administration
\item \verb{[NA(d)]}NA
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\seealso{
Other Core: 
\code{\link{paind80}},
\code{\link{paocc10}},
\code{\link{paocc80}},
\code{\link{paoccindv}},
\code{\link{paoccstatus}},
\code{\link{paocctag}},
\code{\link{papres16}},
\code{\link{papres80}},
\code{\link{pawrkslf}}
}
\concept{Core}
\keyword{variable}
