% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_299.R
\name{painfeel}
\alias{painfeel}
\title{Treatment for pain improve feeling about se}
\source{
General Social Survey https://gss.norc.org
}
\description{
painfeel
}
\details{
Question If you went for treatment because you have PAIN that interferes with your work or other daily activities so that you accomplish less than you would like...
E. Would you definitely expect, probably expect, probably not expect, or definitely not expect to feel better about yourself as a person?
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely expect
\item \verb{[2]} probably expect
\item \verb{[3]} probably not expect
\item \verb{[4]} definitely not expect
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2719/vshow}{GSS Data Explorer page for \code{painfeel}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely expect \tab probably expect \tab probably not expect \tab definitely not expect \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2339 \tab 148 \tab 186 \tab 54 \tab 31 \tab 32 \tab 27 \tab 2817 \cr
   Total \tab 2339 \tab 148 \tab 186 \tab 54 \tab 31 \tab 32 \tab 27 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{downfeel}},
\code{\link{downrely}},
\code{\link{emotcure}},
\code{\link{emotfam}},
\code{\link{emotqol}},
\code{\link{paincure}},
\code{\link{painfam}},
\code{\link{painqol}},
\code{\link{painrely}}
}
\concept{Health Status}
\keyword{variable}
