% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_595.R
\name{paisc681}
\alias{paisc681}
\title{R's father's occupation, 1980 census}
\source{
General Social Survey https://gss.norc.org
}
\description{
paisc681
}
\details{
Question Occupation of respondent's father (International Standard Classification Codes-1968)
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5472/vshow}{GSS Data Explorer page for \code{paisc681}}.
}

\seealso{
Other Occupation: 
\code{\link{isco68}},
\code{\link{isco681}},
\code{\link{paisco68}},
\code{\link{spisco68}}

Other Occupation: 
\code{\link{isco68}},
\code{\link{isco681}},
\code{\link{paisco68}},
\code{\link{spisco68}}
}
\concept{Occupation}
\keyword{variable}
