% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_115.R
\name{pajew}
\alias{pajew}
\title{What sort of jew was r's father}
\source{
General Social Survey https://gss.norc.org
}
\description{
pajew
}
\details{
Question B. Was he Orthodox, Conservative, Reform, or none of these?
}
\section{Values}{

\itemize{
\item \code{1} orthodox
\item \code{2} conservative
\item \code{3} reform
\item \code{4} none of these
\item \code{5} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1066/vshow}{GSS Data Explorer page for \code{pajew}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab orthodox \tab conservative \tab reform \tab none of these \tab don't know \tab no answer \tab Total \cr
   1988 \tab 1447 \tab 9 \tab 7 \tab 8 \tab 2 \tab 1 \tab 7 \tab 1481 \cr
   2008 \tab 1982 \tab 8 \tab 16 \tab 11 \tab 6 \tab - \tab - \tab 2023 \cr
   2012 \tab 1947 \tab 5 \tab 5 \tab 13 \tab 4 \tab - \tab - \tab 1974 \cr
   Total \tab 5376 \tab 22 \tab 28 \tab 32 \tab 12 \tab 1 \tab 7 \tab 5478 \cr
}
}

\seealso{
Other Religion: 
\code{\link{churhsch}},
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{maoth}},
\code{\link{paden}},
\code{\link{pafund}},
\code{\link{paoth}},
\code{\link{parelig}},
\code{\link{sunsch16}}
}
\concept{Religion}
\keyword{variable}
