% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_474.R
\name{pajob}
\alias{pajob}
\title{Type of job held by father when r 16}
\source{
General Social Survey https://gss.norc.org
}
\description{
pajob
}
\details{
Question Here is a list of different types of jobs. Which type did your father have when you were 16. (If your father did not have a job then, please give the job he used to have.)
}
\section{Values}{

\itemize{
\item \verb{[1]} prof or tech
\item \verb{[2]} higher admin
\item \verb{[3]} clerical
\item \verb{[4]} sales
\item \verb{[5]} service
\item \verb{[6]} skilled worker
\item \verb{[7]} semi-skilled worker
\item \verb{[8]} unskilled worker
\item \verb{[9]} farm
\item \verb{[10]} no dad-never had job
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4314/vshow}{GSS Data Explorer page for \code{pajob}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab prof or tech \tab higher admin \tab clerical \tab sales \tab service \tab skilled worker \tab semi-skilled worker \tab unskilled worker \tab farm \tab Total \cr
   1987 \tab 443 \tab 154 \tab 53 \tab 50 \tab 119 \tab 77 \tab 282 \tab 204 \tab 175 \tab 262 \tab 1819 \cr
   Total \tab 443 \tab 154 \tab 53 \tab 50 \tab 119 \tab 77 \tab 282 \tab 204 \tab 175 \tab 262 \tab 1819 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conage}},
\code{\link{firstjob}},
\code{\link{firstslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{occmobil}},
\code{\link{paslf}},
\code{\link{rewrdeff}},
\code{\link{rewrdint}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
