% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_474.R
\name{pajob}
\alias{pajob}
\title{Type of job held by father when r 16}
\source{
General Social Survey https://gss.norc.org
}
\description{
pajob
}
\details{
Question Here is a list of different types of jobs. Which type did your father have when you were 16. (If your father did not have a job then, please give the job he used to have.)
}
\section{Values}{

\itemize{
\item \code{1} prof or tech
\item \code{2} higher admin
\item \code{3} clerical
\item \code{4} sales
\item \code{5} service
\item \code{6} skilled worker
\item \code{7} semi-skilled worker
\item \code{8} unskilled worker
\item \code{9} farm
\item \code{10} no dad-never had job
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4314/vshow}{GSS Data Explorer page for \code{pajob}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conage}},
\code{\link{firstjob}},
\code{\link{firstslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{occmobil}},
\code{\link{paslf}},
\code{\link{rewrdeff}},
\code{\link{rewrdint}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
