% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_662.R
\name{pandmet}
\alias{pandmet}
\title{Change in socialization over covid pandemic}
\source{
General Social Survey https://gss.norc.org
}
\description{
pandmet
}
\details{
Question Thinking about how often you met with your extended family and friends in person before the COVID-19 pandemic, would you say you now meet them in person more often, less often, or about the same?

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \code{1} much less often
\item \code{2} a little less often
\item \code{3} about the same
\item \code{4} a little more often
\item \code{5} much more often
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7955/vshow}{GSS Data Explorer page for \code{pandmet}}.
}

\seealso{
Other GSS Next: 
\code{\link{abbelief}},
\code{\link{abgender}},
\code{\link{abhelp1}},
\code{\link{abhelp2}},
\code{\link{abhelp3}},
\code{\link{abhelp4}},
\code{\link{abmoral}},
\code{\link{advfront}},
\code{\link{biokids}},
\code{\link{blkdsrv}},
\code{\link{blkmblty}},
\code{\link{blktry}},
\code{\link{brv16}},
\code{\link{brv16bronum}},
\code{\link{brv16cuznum}},
\code{\link{brv16dadnum}},
\code{\link{brv16filnum}},
\code{\link{brv16frndnum}},
\code{\link{brv16gmanum}},
\code{\link{brv16gpanum}},
\code{\link{brv16grand}},
\code{\link{brv16milnum}},
\code{\link{brv16momnum}},
\code{\link{brv16oth}},
\code{\link{brv16othnum}},
\code{\link{brv16par}},
\code{\link{brv16partnum}},
\code{\link{brv16sib}},
\code{\link{brv16silnum}},
\code{\link{brv16sisnum}},
\code{\link{brv16sp}},
\code{\link{brv16spnum}},
\code{\link{brv5}},
\code{\link{brv5bronum}},
\code{\link{brv5child}},
\code{\link{brv5chinum}},
\code{\link{brv5cowknum}},
\code{\link{brv5cuznum}},
\code{\link{brv5dadnum}},
\code{\link{brv5daunum}},
\code{\link{brv5filnum}},
\code{\link{brv5frndnum}},
\code{\link{brv5gmanum}},
\code{\link{brv5gpanum}},
\code{\link{brv5grand}},
\code{\link{brv5milnum}},
\code{\link{brv5momnum}},
\code{\link{brv5oth}},
\code{\link{brv5othnum}},
\code{\link{brv5par}},
\code{\link{brv5partnum}},
\code{\link{brv5sib}},
\code{\link{brv5silnum}},
\code{\link{brv5sisnum}},
\code{\link{brv5sonnum}},
\code{\link{brv5sp}},
\code{\link{brv5spnum}},
\code{\link{confedvacy}},
\code{\link{conpharvacy}},
\code{\link{covemply}},
\code{\link{covid12}},
\code{\link{femself}},
\code{\link{firstkidsex}},
\code{\link{fluvax}},
\code{\link{fund_next}},
\code{\link{gestate}},
\code{\link{malekids}},
\code{\link{mascself}},
\code{\link{nexteligible}},
\code{\link{nextgen}},
\code{\link{nonbinkids}},
\code{\link{pandinc}},
\code{\link{scientbe}},
\code{\link{scientgo}},
\code{\link{scienthe}},
\code{\link{toofast}},
\code{\link{vaxhstncy}},
\code{\link{vaxkids}},
\code{\link{vaxsafe}}
}
\concept{GSS Next}
\keyword{variable}
