% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_06.R
\name{paocc16}
\alias{paocc16}
\title{Father's census occupation code (1970)}
\source{
General Social Survey https://gss.norc.org
}
\description{
paocc16
}
\details{
Question A. What kind of work did your father (FATHER SUBSTITUTE) normally do while you were growing up? That is, what was his job called?
B. What did he actually do in that job? Tell me, what were some of his main duties?
C. What kind of place did he work for?
D. What did they (make/do)?
E. Was he self-employed or did he work for someone else?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\seealso{
Other Core: 
\code{\link{coind10}},
\code{\link{coocc10}},
\code{\link{spind10}},
\code{\link{spindstatus}},
\code{\link{spindtag}},
\code{\link{spoccstatus}},
\code{\link{spocctag}},
\code{\link{sppres10}},
\code{\link{sppres105plus}}
}
\concept{Core}
\keyword{variable}
