% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_07.R
\name{paocc80}
\alias{paocc80}
\title{Father's census occupation code (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
paocc80
}
\details{
Question Father's Occupation, Prestige, and Industry
}
\section{Values}{

\itemize{
\item \verb{[3]} legislators
\item \verb{[4]} chief executives and general admin., public admin
\item \verb{[5]} administrators and officials, public administration
\item \verb{[6]} administrators, protective service
\item \verb{[7]} financial managers
\item \verb{[8]} personnel and labor relations managers
\item \verb{[9]} purchasing managers
\item \verb{[13]} managers, marketing, advertising, and public relations
\item \verb{[14]} administrators, education and related fields
\item \verb{[15]} managers, medicine and health
\item \verb{[16]} managers, properties and real estate
\item \verb{[17]} postmasters and mail superintendents
\item \verb{[18]} funeral directors
\item \verb{[19]} managers and administrators, n.e.c.
\item \verb{[23]} accountants and auditors
\item \verb{[24]} underwriters
\item \verb{[25]} other financial officers
\item \verb{[26]} management analysts
\item \verb{[27]} personnel, training, and labor relations specialists
\item \verb{[28]} purchasing agents and buyers, farm products
\item \verb{[29]} buyers, wholesale and retail trade except farm products
\item \verb{[33]} purchasing agents and buyers
\item \verb{[34]} business and promotion agents
\item \verb{[35]} construction inspectors
\item \verb{[36]} inspectors/compliance officers, except construction
\item \verb{[37]} management related occupations, n.e.c.
\item \verb{[43]} architects
\item \verb{[44]} aerospace engineers
\item \verb{[45]} metallurgical and materials engineers
\item \verb{[46]} mining engineers
\item \verb{[47]} petroleum engineers
\item \verb{[48]} chemical engineers
\item \verb{[49]} nuclear engineers
\item \verb{[53]} civil engineers
\item \verb{[54]} agricultural engineers
\item \verb{[55]} electrical and electronic engineers
\item \verb{[56]} industrial engineers
\item \verb{[57]} mechanical engineers
\item \verb{[58]} marine and naval architects
\item \verb{[59]} engineers, n.e.c.
\item \verb{[63]} surveyors and mapping scientists
\item \verb{[64]} computer systems analysts and scientists
\item \verb{[65]} operations and systems researchers and analysts
\item \verb{[66]} actuaries
\item \verb{[67]} statisticians
\item \verb{[68]} mathematical scientists, n.e.c.
\item \verb{[69]} physicists and astronomers
\item \verb{[73]} chemists, except biochemists
\item \verb{[74]} atmospheric and space scientists
\item \verb{[75]} geologists and geodesists
\item \verb{[76]} physical scientists, n.e.c.
\item \verb{[77]} agricultural and food scientists
\item \verb{[78]} biological and life scientists
\item \verb{[79]} forestry and conservation scientists
\item \verb{[83]} medical scientists
\item \verb{[84]} physicians
\item \verb{[85]} dentists
\item \verb{[86]} veterinarians
\item \verb{[87]} optometrists
\item \verb{[88]} podiatrists
\item \verb{[89]} health diagnosing practitioners, n.e.c.
\item \verb{[95]} registered nurses
\item \verb{[96]} pharmacists
\item \verb{[97]} dietitians
\item \verb{[98]} inhalation therapists
\item \verb{[99]} occupational therapists
\item \verb{[103]} physical therapists
\item \verb{[104]} speech therapists
\item \verb{[105]} therapists, n.e.c.
\item \verb{[106]} physicians' assistants
\item \verb{[113]} earth, environmental, and marine science teachers
\item \verb{[114]} biological science teachers
\item \verb{[115]} chemistry teachers
\item \verb{[116]} physics teachers
\item \verb{[117]} natural science teachers, n.e.c.
\item \verb{[118]} psychology teachers
\item \verb{[119]} economics teachers
\item \verb{[123]} history teachers
\item \verb{[124]} political science teachers
\item \verb{[125]} sociology teachers
\item \verb{[126]} social science teachers, n.e.c.
\item \verb{[127]} engineering teachers
\item \verb{[128]} mathematical science teachers
\item \verb{[129]} computer science teachers
\item \verb{[133]} medical science teachers
\item \verb{[134]} health specialties teachers
\item \verb{[135]} business, commerce, and marketing teachers
\item \verb{[136]} agriculture and forestry teachers
\item \verb{[137]} art, drama, and music teachers
\item \verb{[138]} physical education teachers
\item \verb{[139]} education teachers
\item \verb{[143]} english teachers
\item \verb{[144]} foreign language teachers
\item \verb{[145]} law teachers
\item \verb{[146]} social work teachers
\item \verb{[147]} theology teachers
\item \verb{[148]} trade and industrial teachers
\item \verb{[149]} home economics teachers
\item \verb{[153]} teachers, postsecondary, n.e.c.
\item \verb{[154]} postsecondary teachers, subject not specified
\item \verb{[155]} teachers, prekindergarten and kindergarten
\item \verb{[156]} teachers, elementary school
\item \verb{[157]} teachers, secondary school
\item \verb{[158]} teachers, special education
\item \verb{[159]} teachers, n.e.c.
\item \verb{[163]} counselors, educational and vocational
\item \verb{[164]} librarians
\item \verb{[165]} archivists and curators
\item \verb{[166]} economists
\item \verb{[167]} psychologists
\item \verb{[168]} sociologists
\item \verb{[169]} social scientists, n.e.c.
\item \verb{[173]} urban planners
\item \verb{[174]} social workers
\item \verb{[175]} recreation workers
\item \verb{[176]} clergy
\item \verb{[177]} religious workers, n.e.c.
\item \verb{[178]} lawyers
\item \verb{[179]} judges
\item \verb{[183]} authors
\item \verb{[184]} technical writers
\item \verb{[185]} designers
\item \verb{[186]} musicians and composers
\item \verb{[187]} actors and directors
\item \verb{[188]} painters, sculptors, craft-artists, and printmakers
\item \verb{[189]} photographers
\item \verb{[193]} dancers
\item \verb{[194]} artists, performers, and related workers, n.e.c.
\item \verb{[195]} editors and reporters
\item \verb{[197]} public relations specialists
\item \verb{[198]} announcers
\item \verb{[199]} athletes
\item \verb{[203]} clinical laboratory technologists and technicians
\item \verb{[204]} dental hygienists
\item \verb{[205]} health record technologists and technicians
\item \verb{[206]} radiologic technicians
\item \verb{[207]} licensed practical nurses
\item \verb{[208]} health technologists and technicians, n.e.c.
\item \verb{[213]} electrical and electronic technicians
\item \verb{[214]} industrial engineering technicians
\item \verb{[215]} mechanical engineering technicians
\item \verb{[216]} engineering technicians, n.e.c.
\item \verb{[217]} drafting occupations
\item \verb{[218]} surveying and mapping technicians
\item \verb{[223]} biological technicians
\item \verb{[224]} chemical technicians
\item \verb{[225]} science technicians, n.e.c.
\item \verb{[226]} airplane pilots and navigators
\item \verb{[227]} air traffic controllers
\item \verb{[228]} broadcast equipment operators
\item \verb{[229]} computer programmers
\item \verb{[233]} tool programmers, numerical control
\item \verb{[234]} legal assistants
\item \verb{[235]} technicians, n.e.c.
\item \verb{[243]} supervisors and proprietors, sales occupations
\item \verb{[253]} insurance sales occupations
\item \verb{[254]} real estate sales occupations
\item \verb{[255]} securities and financial services sales occupations
\item \verb{[256]} advertising and related sales occupations
\item \verb{[257]} sales occupations, other business services
\item \verb{[258]} sales engineers
\item \verb{[259]} sales rep., mining, manufacturing, and wholesale
\item \verb{[263]} sales workers, motor vehicles and boats
\item \verb{[264]} sales workers, apparel
\item \verb{[265]} sales workers, shoes
\item \verb{[266]} sales workers, furniture and home furnishings
\item \verb{[267]} sales workers, radio, tv, hi-fi, and appliances
\item \verb{[268]} sales workers, hardware and building supplies
\item \verb{[269]} sales workers, parts
\item \verb{[274]} sales workers, other commodities
\item \verb{[275]} sales counter clerks
\item \verb{[276]} cashiers
\item \verb{[277]} street and door-to-door sales workers
\item \verb{[278]} news vendors
\item \verb{[283]} demonstrators, promoters and models, sales
\item \verb{[284]} auctioneers
\item \verb{[285]} sales support occupations, n.e.c.
\item \verb{[303]} supervisors, general office
\item \verb{[304]} supervisors, computer equipment operators
\item \verb{[305]} supervisors, financial records processing
\item \verb{[306]} chief communications operators
\item \verb{[307]} supervisors, distribution, scheduling, and adjusting clerks
\item \verb{[308]} computer operators
\item \verb{[309]} peripheral equipment operators
\item \verb{[313]} secretaries
\item \verb{[314]} stenographers
\item \verb{[315]} typists
\item \verb{[316]} interviewers
\item \verb{[317]} hotel clerks
\item \verb{[318]} transportation ticket and reservation agents
\item \verb{[319]} receptionists
\item \verb{[323]} information clerks, n.e.c.
\item \verb{[325]} classified-ad clerks
\item \verb{[326]} correspondence clerks
\item \verb{[327]} order clerks
\item \verb{[328]} personnel clerks, except payroll and timekeeping
\item \verb{[329]} library clerks
\item \verb{[335]} file clerks
\item \verb{[336]} records clerks
\item \verb{[337]} bookkeepers, accounting and auditing clerks
\item \verb{[338]} payroll and timekeeping clerks
\item \verb{[339]} billing clerks
\item \verb{[343]} cost and rate clerks
\item \verb{[344]} billing, posting, and calculating machine operators
\item \verb{[345]} duplicating machine operators
\item \verb{[346]} mail preparing and paper handling machine operators
\item \verb{[347]} office machine operators, n.e.c.
\item \verb{[348]} telephone operators
\item \verb{[349]} telegraphers
\item \verb{[353]} communications equipment operators, n.e.c.
\item \verb{[354]} postal clerks, excluding mail carriers
\item \verb{[355]} mail carriers, postal service
\item \verb{[356]} mail clerks, excluding postal service
\item \verb{[357]} messengers
\item \verb{[359]} dispatchers
\item \verb{[363]} production coordinators
\item \verb{[364]} traffic, shipping and receiving clerks
\item \verb{[365]} stock and inventory clerks
\item \verb{[366]} meter readers
\item \verb{[368]} weighers, measurers, and checkers
\item \verb{[369]} samplers
\item \verb{[373]} expediters
\item \verb{[374]} material recording, scheduling and distributing clerks, n.e.c.
\item \verb{[375]} insurance adjusters, examiners, and investigators
\item \verb{[376]} investigators and adjusters, except insurance
\item \verb{[377]} eligibility clerks, social welfare
\item \verb{[378]} bill and account collectors
\item \verb{[379]} general office clerks
\item \verb{[383]} bank tellers
\item \verb{[384]} proofreaders
\item \verb{[385]} data-entry keyers
\item \verb{[386]} statistical clerks
\item \verb{[387]} teachers' aides
\item \verb{[389]} administrative support occupations, n.e.c.
\item \verb{[403]} launderers and ironers
\item \verb{[404]} cooks, private household
\item \verb{[405]} housekeepers and butlers
\item \verb{[406]} child care workers, private household
\item \verb{[407]} private household cleaners and servants
\item \verb{[413]} supervisors, firefighting and fire prevention occupations
\item \verb{[414]} supervisors, police and detectives
\item \verb{[415]} supervisors, guards
\item \verb{[416]} fire inspection and fire prevention occupations
\item \verb{[417]} firefighting occupations
\item \verb{[418]} police and detectives, public service
\item \verb{[423]} sheriffs, bailiffs, and other law enforcement officers
\item \verb{[424]} correctional institution officers
\item \verb{[425]} crossing guards
\item \verb{[426]} guards and police, excluding public service
\item \verb{[427]} protective service occupations
\item \verb{[430]} former member of the armed forces
\item \verb{[431]} current member of the armed forces
\item \verb{[433]} supervisors, food preparation and service occupations
\item \verb{[434]} bartenders
\item \verb{[435]} waiters and waitresses
\item \verb{[436]} cooks, except short order
\item \verb{[437]} short-order cooks
\item \verb{[438]} food counter, fountain and related occupations
\item \verb{[439]} kitchen workers, food preparation
\item \verb{[443]} waiters'/waitresses' assistants
\item \verb{[444]} miscellaneous food preparation occupations
\item \verb{[445]} dental assistants
\item \verb{[446]} health aids, except nursing
\item \verb{[447]} nursing aides, orderlies and attendants
\item \verb{[448]} supervisors, cleaning and building service workers
\item \verb{[449]} maids and housemen
\item \verb{[453]} janitors and cleaners
\item \verb{[454]} elevator operators
\item \verb{[455]} pest control occupations
\item \verb{[456]} supervisors, personal service occupations
\item \verb{[457]} barbers
\item \verb{[458]} hairdressers and cosmetologists
\item \verb{[459]} attendants, amusement and recreation facilities
\item \verb{[463]} guides
\item \verb{[464]} ushers
\item \verb{[465]} public transportation attendants
\item \verb{[466]} baggage porters and bellhops
\item \verb{[467]} welfare service aides
\item \verb{[468]} child care workers, except private household
\item \verb{[469]} personal service occupations, n.e.c.
\item \verb{[473]} farmers, except horticultural
\item \verb{[474]} horticultural specialty farmers
\item \verb{[475]} managers, farms, except horticultural
\item \verb{[476]} managers, horticultural specialty farms
\item \verb{[477]} supervisors, farm workers
\item \verb{[479]} farm workers
\item \verb{[483]} marine life cultivation workers
\item \verb{[484]} nursery workers
\item \verb{[485]} supervisors, related agricultural occupations
\item \verb{[486]} groundskeepers and gardeners, except farm
\item \verb{[487]} animal caretakers, except farm
\item \verb{[488]} graders and sorters, agricultural products
\item \verb{[489]} inspectors, agricultural products
\item \verb{[494]} supervisors, forestry and logging workers
\item \verb{[495]} forestry workers, except logging
\item \verb{[496]} timber cutting and logging occupations
\item \verb{[497]} captains and other officers, fishing vessels
\item \verb{[498]} fishers
\item \verb{[499]} hunters and trappers
\item \verb{[503]} supervisors, mechanics and repairers
\item \verb{[505]} automobile mechanics, except apprentices
\item \verb{[506]} automobile mechanic apprentices
\item \verb{[507]} bus, truck, and stationary engine mechanics
\item \verb{[508]} aircraft engine mechanics
\item \verb{[509]} small engine repairers
\item \verb{[514]} automobile body and related repairers
\item \verb{[515]} aircraft mechanics, excluding engine
\item \verb{[516]} heavy equipment mechanics
\item \verb{[517]} farm equipment mechanics
\item \verb{[518]} industrial machinery repairers
\item \verb{[519]} machinery maintenance occupations
\item \verb{[523]} electronic repairers, communications and industrial equipment
\item \verb{[525]} data processing equipment repairers
\item \verb{[526]} household appliance and power tool repairers
\item \verb{[527]} telephone line installers and repairers
\item \verb{[529]} telephone installers and repairers
\item \verb{[533]} misc. electrical and electronic equipment repairers
\item \verb{[534]} heating, air conditioning, and refrigeration mechanics
\item \verb{[535]} camera, watch, and musical instrument repairers
\item \verb{[536]} locksmiths and safe repairers
\item \verb{[538]} office machine repairers
\item \verb{[539]} mechanical controls and valve repairers
\item \verb{[543]} elevator installers and repairers
\item \verb{[544]} millwrights
\item \verb{[547]} specified mechanics and repairers, n.e.c.
\item \verb{[549]} not specified mechanics and repairers
\item \verb{[553]} supervisors, brickmasons, stonemasons, and title setters
\item \verb{[554]} supervisors, carpenters and related work
\item \verb{[555]} supervisors, electricians and power transmission installers
\item \verb{[556]} supervisors, painters, paperhangers, and plasterers
\item \verb{[557]} supervisors, plumbers, pipefitters, and steamfitters
\item \verb{[558]} supervisors, n.e.c.
\item \verb{[563]} brickmasons and stonemasons, except apprentices
\item \verb{[564]} brickmasons and stonemasons apprentices
\item \verb{[565]} tile setters, hard and soft
\item \verb{[566]} carpet installers
\item \verb{[567]} carpenters, except apprentices
\item \verb{[569]} carpenter apprentices
\item \verb{[573]} drywall installers
\item \verb{[575]} electricians, except apprentices
\item \verb{[576]} electrician apprentices
\item \verb{[577]} electrical power installers and repairers
\item \verb{[579]} painters, construction and maintenance
\item \verb{[583]} paperhangers
\item \verb{[584]} plasterers
\item \verb{[585]} plumbers, pipefitters, and steamfitters, except apprentices
\item \verb{[587]} plumber, pipefitter, and steamfitter apprentices
\item \verb{[588]} concrete and terrazzo finishers
\item \verb{[589]} glaziers
\item \verb{[593]} insulation workers
\item \verb{[594]} paving, surfacing, and tamping equipment operators
\item \verb{[595]} roofers
\item \verb{[596]} sheetmetal duct installers
\item \verb{[597]} structural metal workers
\item \verb{[598]} drillers, earth
\item \verb{[599]} construction trades, n.e.c.
\item \verb{[613]} supervisors, extractive occupations
\item \verb{[614]} drillers, oil well
\item \verb{[615]} explosives workers
\item \verb{[616]} mining machine operators
\item \verb{[617]} mining occupations, n.e.c.
\item \verb{[633]} supervisors, production occupations
\item \verb{[634]} tool and die makers, except apprentices
\item \verb{[635]} tool and die maker apprentices
\item \verb{[636]} precision assemblers, metal
\item \verb{[637]} machinists, except apprentices
\item \verb{[639]} machinist apprentices
\item \verb{[643]} boilermakers
\item \verb{[644]} precision grinders, fitters, and tool sharpeners
\item \verb{[645]} patternmakers and model makers, metal
\item \verb{[646]} lay-out workers
\item \verb{[647]} precious stones and metals workers
\item \verb{[649]} engravers, metal
\item \verb{[653]} sheet metal workers, except apprentices
\item \verb{[654]} sheet metal worker, apprentices
\item \verb{[655]} miscellaneous precision metal workers
\item \verb{[656]} patternmakers and model makers, wood
\item \verb{[657]} cabinet makers and bench carpenters
\item \verb{[658]} furniture and wood finishers
\item \verb{[659]} miscellaneous precision woodworkers
\item \verb{[666]} dressmakers
\item \verb{[667]} tailors
\item \verb{[668]} upholsterers
\item \verb{[669]} shoe repairers
\item \verb{[673]} apparel and fabric patternmakers
\item \verb{[674]} miscellaneous precision apparel and fabric workers
\item \verb{[675]} hand molders and shapers, except jewelers
\item \verb{[676]} patternmakers, lay-out workers, and cutters
\item \verb{[677]} optical goods workers
\item \verb{[678]} dental laboratory and medical appliance technicians
\item \verb{[679]} bookbinders
\item \verb{[683]} electrical and electronic equipment assemblers
\item \verb{[684]} miscellaneous precision workers, n.e.c.
\item \verb{[686]} butchers and meat cutters
\item \verb{[687]} bakers
\item \verb{[688]} food batchmakers
\item \verb{[689]} inspectors, testers, and graders
\item \verb{[693]} adjusters and calibrators
\item \verb{[694]} water and sewage treatment plant operators
\item \verb{[695]} power plant operators
\item \verb{[696]} stationary engineers
\item \verb{[699]} miscellaneous plant and system operators
\item \verb{[703]} lathe and turning machine set-up operators
\item \verb{[704]} lathe and turning machine operators
\item \verb{[705]} milling and planing machine operators
\item \verb{[706]} punching and stamping press machine operators
\item \verb{[707]} rolling machine operators
\item \verb{[708]} drilling and boring machine operators
\item \verb{[709]} grinding, abrading, buffing, and polishing machine operators
\item \verb{[713]} forging machine operators
\item \verb{[714]} numerical control machine operators
\item \verb{[715]} miscellaneous metal, plastic, stone, and glass working machine operators
\item \verb{[717]} fabricating machine operators, n.e.c.
\item \verb{[719]} molding and casting machine operators
\item \verb{[723]} metal plating machine operators
\item \verb{[724]} heat treating equipment operators
\item \verb{[725]} miscellaneous metal and plastic processing machine operators
\item \verb{[726]} wood lathe, routing and planing machine operators
\item \verb{[727]} sawing machine operators
\item \verb{[728]} shaping and joining machine operators
\item \verb{[729]} nailing and tacking machine operators
\item \verb{[733]} miscellaneous woodworking machine operators
\item \verb{[734]} printing machine operators
\item \verb{[735]} photoengravers and lithographers
\item \verb{[736]} typesetters and compositors
\item \verb{[737]} miscellaneous printing machine operators
\item \verb{[738]} winding and twisting machine operators
\item \verb{[739]} knitting, looping, taping, and weaving machine operators
\item \verb{[743]} textile cutting machine operators
\item \verb{[744]} textile sewing machine operators
\item \verb{[745]} shoe machine operators
\item \verb{[747]} pressing machine operators
\item \verb{[748]} laundering and dry cleaning machine operators
\item \verb{[749]} miscellaneous textile machine operators
\item \verb{[753]} cementing and gluing machine operators
\item \verb{[754]} packaging and filling machine operators
\item \verb{[755]} extruding and forming machine operators
\item \verb{[756]} mixing and blending machine operators
\item \verb{[757]} separating, filtering, and clarifying machine operators
\item \verb{[758]} compressing and compacting machine operators
\item \verb{[759]} painting and paint spraying machine operators
\item \verb{[763]} roasting and baking machine operators, food
\item \verb{[764]} washing, cleaning, and pickling machine operators
\item \verb{[765]} folding machine operators
\item \verb{[766]} furnace, kiln, and oven operators, except food
\item \verb{[768]} crushing and grinding machine operators
\item \verb{[769]} slicing and cutting machine operators
\item \verb{[773]} motion picture projectionists
\item \verb{[774]} photographic process machine operators
\item \verb{[777]} miscellaneous and not specified machine operators, n.e.c.
\item \verb{[779]} machine operators, not specified
\item \verb{[783]} welders and cutters
\item \verb{[784]} solderers and blazers
\item \verb{[785]} assemblers
\item \verb{[786]} hand cutting and trimming occupations
\item \verb{[787]} hand molding, casting, and forming occupations
\item \verb{[789]} hand painting,coating, and decorating occupations
\item \verb{[793]} hand engraving and printing occupations
\item \verb{[794]} hand grinding and polishing occupations
\item \verb{[795]} miscellaneous hand working occupations
\item \verb{[796]} production inspectors, checkers, and examiners
\item \verb{[797]} production testers
\item \verb{[798]} production samplers and weighers
\item \verb{[799]} graders and sorters, except agricultural
\item \verb{[803]} supervisors, motor vehicle operators
\item \verb{[804]} truck drivers, heavy
\item \verb{[805]} truck drivers, light
\item \verb{[806]} driver-sales workers
\item \verb{[808]} bus drivers
\item \verb{[809]} taxicab drivers and chauffeurs
\item \verb{[813]} parking lot attendants
\item \verb{[814]} motor transportation occupations, n.e.c.
\item \verb{[823]} railroad conductors and yardmasters
\item \verb{[824]} locomotive operating occupations
\item \verb{[825]} railroad brake, signal, and switch operators
\item \verb{[826]} rail vehicle operators, n.e.c.
\item \verb{[828]} ship captains and mates, except fishing boats
\item \verb{[829]} sailors and deckhands
\item \verb{[833]} marine engineers
\item \verb{[834]} bridge, lock and lighthouse tenders
\item \verb{[843]} supervisors, material moving equipment operators
\item \verb{[844]} operating engineers
\item \verb{[845]} longshore equipment operators
\item \verb{[848]} hoist and winch operators
\item \verb{[849]} crane and tower operators
\item \verb{[853]} excavating and loading machine operators
\item \verb{[855]} grader, dozer, and scraper operators
\item \verb{[856]} industrial truck and tractor equipment operators
\item \verb{[859]} miscellaneous material moving equipment operators
\item \verb{[863]} handlers, equipment cleaners, helpers, and laborers, n.e.c.
\item \verb{[864]} helpers, mechanics and repairers
\item \verb{[865]} helpers, construction trades
\item \verb{[866]} helpers, surveyor
\item \verb{[867]} helpers, extractive occupations
\item \verb{[869]} construction laborers
\item \verb{[873]} production helpers
\item \verb{[875]} garbage collectors
\item \verb{[876]} stevedores
\item \verb{[877]} stock handlers and baggers
\item \verb{[878]} machine feeders and offbearers
\item \verb{[883]} freight, stock, and material handlers, n.e.c.
\item \verb{[885]} garage and service station related occupations
\item \verb{[887]} vehicle washers and equipment cleaners
\item \verb{[888]} hand packers and packagers
\item \verb{[889]} laborers, except construction
\item \verb{[998]} dk
\item \verb{[999]} na
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010 \cr
}
}

\seealso{
Other Core: 
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{paocc10}},
\code{\link{paoccindv}},
\code{\link{paoccstatus}},
\code{\link{paocctag}},
\code{\link{papres16}},
\code{\link{papres80}},
\code{\link{pawrkslf}}
}
\concept{Core}
\keyword{variable}
