% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_115.R
\name{paoth}
\alias{paoth}
\title{Other protestant denomination for father}
\source{
General Social Survey https://gss.norc.org
}
\description{
paoth
}
\details{
Question A. What specific denomination is that, if any?
}
\section{Values}{

\itemize{
\item \code{149} polish catholic
\item \code{196} empowerment temple
\item \code{197} grace independent baptist church
\item \code{198} new life
\item \code{199} follower of christ
\item \code{201} pathways christian church
\item \code{204} renia de deus
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1065/vshow}{GSS Data Explorer page for \code{paoth}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 111 \tab 116 \tab 12 \tab 120 \tab 132 \tab 18 \tab 20 \tab 23 \tab 3 \tab 31 \tab 32 \tab 33 \tab 35 \tab 36 \tab 37 \tab 40 \tab 43 \tab 44 \tab 45 \tab 46 \tab 51 \tab 54 \tab 56 \tab 57 \tab 58 \tab 59 \tab 60 \tab 61 \tab 64 \tab 65 \tab 68 \tab 69 \tab 7 \tab 70 \tab 71 \tab 74 \tab 8 \tab 80 \tab 81 \tab 83 \tab 96 \tab no answer \tab Total \cr
   1988 \tab 1350 \tab 1 \tab 1 \tab 8 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 6 \tab 2 \tab 1 \tab 9 \tab 4 \tab 1 \tab 10 \tab 1 \tab 5 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 14 \tab 2 \tab 1 \tab 8 \tab 2 \tab 7 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 10 \tab 1481 \cr
   Total \tab 1350 \tab 1 \tab 1 \tab 8 \tab 1 \tab 1 \tab 1 \tab 3 \tab 3 \tab 1 \tab 6 \tab 2 \tab 1 \tab 9 \tab 4 \tab 1 \tab 10 \tab 1 \tab 5 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 2 \tab 14 \tab 2 \tab 1 \tab 8 \tab 2 \tab 7 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 10 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{churhsch}},
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{maoth}},
\code{\link{paden}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{parelig}},
\code{\link{sunsch16}}
}
\concept{Religion}
\keyword{variable}
