% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_08.R
\name{papres105plus}
\alias{papres105plus}
\title{Father's occupational prestige score using threshold method (2010)}
\source{
General Social Survey https://gss.norc.org
}
\description{
papres105plus
}
\details{
Question PAPRES105PLUS is based on the 2010 Census occupation classification. This prestige score is calculated using an alternative method, based on the percentage of ratings that is greater than or equal to a threshold (rating five). The effect of individual raters is also removed using hierarchical linear modeling (HLM).
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/- \tab full \cr
   1973 \tab -/-/- \tab full \cr
   1974 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1976 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5711/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 58 \tab 59 \tab 6 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 69 \tab 72 \tab 73 \tab 75 \tab 76 \tab 78 \tab 8 \tab 80 \tab 81 \tab 82 \tab 83 \tab 84 \tab 85 \tab 86 \tab 87 \tab 88 \tab 89 \tab 90 \tab 91 \tab 92 \tab 94 \tab 95 \tab iap \tab 43 \tab 57 \tab 68 \tab 70 \tab 71 \tab 77 \tab 9 \tab 93 \tab 30 \tab 7 \tab 74 \tab 97 \tab 5 \tab 79 \tab 67 \tab not available in this release \tab Total \cr
   1972 \tab 4 \tab 7 \tab 10 \tab 4 \tab 3 \tab 18 \tab 23 \tab 7 \tab 71 \tab 31 \tab 3 \tab 14 \tab 3 \tab 23 \tab 9 \tab 3 \tab 32 \tab 19 \tab 12 \tab 59 \tab 5 \tab 3 \tab 12 \tab 5 \tab 7 \tab 11 \tab 23 \tab 18 \tab 17 \tab 10 \tab 1 \tab 43 \tab 5 \tab 7 \tab 17 \tab 3 \tab 7 \tab 50 \tab 41 \tab 4 \tab 1 \tab 17 \tab 12 \tab 12 \tab 4 \tab 305 \tab 1 \tab 1 \tab 4 \tab 14 \tab 30 \tab 12 \tab 3 \tab 12 \tab 2 \tab 6 \tab 2 \tab 2 \tab 4 \tab 7 \tab 10 \tab 2 \tab 10 \tab 1 \tab 2 \tab 1 \tab 6 \tab 13 \tab 11 \tab 2 \tab 4 \tab 3 \tab 7 \tab 11 \tab 3 \tab 18 \tab 8 \tab 406 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1 \tab 8 \tab 2 \tab 3 \tab 1 \tab 12 \tab 27 \tab 2 \tab 101 \tab 17 \tab 5 \tab 22 \tab 5 \tab 21 \tab 16 \tab 11 \tab 27 \tab 13 \tab 28 \tab 65 \tab 4 \tab 4 \tab 10 \tab 5 \tab 11 \tab 18 \tab 19 \tab 11 \tab 38 \tab 7 \tab 2 \tab 47 \tab 7 \tab 9 \tab 10 \tab 5 \tab 8 \tab 52 \tab 32 \tab 6 \tab 2 \tab 12 \tab 17 \tab 24 \tab 4 \tab 240 \tab 1 \tab 1 \tab 3 \tab 9 \tab 40 \tab 26 \tab 1 \tab 11 \tab 2 \tab 5 \tab - \tab - \tab 9 \tab 4 \tab 9 \tab 2 \tab 14 \tab 3 \tab 2 \tab 1 \tab 5 \tab 16 \tab 12 \tab 3 \tab - \tab 4 \tab 14 \tab 11 \tab 4 \tab 17 \tab 9 \tab 254 \tab 13 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 8 \tab 8 \tab - \tab 7 \tab 12 \tab 18 \tab 4 \tab 84 \tab 14 \tab 10 \tab 9 \tab 5 \tab 18 \tab 17 \tab 7 \tab 16 \tab 18 \tab 26 \tab 51 \tab 5 \tab 13 \tab 12 \tab 9 \tab 12 \tab 11 \tab 15 \tab 8 \tab 28 \tab 14 \tab - \tab 37 \tab 3 \tab 8 \tab 15 \tab 5 \tab 5 \tab 56 \tab 54 \tab 10 \tab 4 \tab 19 \tab 14 \tab 23 \tab 2 \tab 237 \tab 1 \tab 2 \tab 2 \tab 11 \tab 24 \tab 15 \tab 4 \tab 6 \tab 3 \tab 7 \tab 5 \tab 9 \tab 5 \tab 8 \tab 11 \tab 7 \tab 3 \tab 3 \tab 1 \tab - \tab 4 \tab 21 \tab 6 \tab 2 \tab 3 \tab 3 \tab 12 \tab 13 \tab 4 \tab 20 \tab 4 \tab 294 \tab 8 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1 \tab 7 \tab 4 \tab 3 \tab 3 \tab 12 \tab 21 \tab 3 \tab 43 \tab 29 \tab 7 \tab 19 \tab 7 \tab 21 \tab 13 \tab 6 \tab 22 \tab 28 \tab 15 \tab 74 \tab 5 \tab 2 \tab 4 \tab 9 \tab 16 \tab 15 \tab 17 \tab 12 \tab 30 \tab 9 \tab - \tab 46 \tab 6 \tab 7 \tab 15 \tab 4 \tab 4 \tab 50 \tab 40 \tab 5 \tab 2 \tab 19 \tab 17 \tab 7 \tab 1 \tab 270 \tab 5 \tab 2 \tab 8 \tab 11 \tab 31 \tab 18 \tab 1 \tab 11 \tab 4 \tab 11 \tab 1 \tab 3 \tab 7 \tab 8 \tab 7 \tab 2 \tab 7 \tab 8 \tab 1 \tab 1 \tab 7 \tab 14 \tab 9 \tab 3 \tab 1 \tab 2 \tab 10 \tab 7 \tab 1 \tab 18 \tab 11 \tab 289 \tab 14 \tab - \tab 2 \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1 \tab 4 \tab 8 \tab 2 \tab 2 \tab 19 \tab 37 \tab 4 \tab 53 \tab 19 \tab 12 \tab 12 \tab 5 \tab 28 \tab 10 \tab 7 \tab 26 \tab 15 \tab 35 \tab 69 \tab 7 \tab 5 \tab 8 \tab 11 \tab 6 \tab 9 \tab 26 \tab 8 \tab 41 \tab 6 \tab 1 \tab 41 \tab 5 \tab 9 \tab 10 \tab 8 \tab 3 \tab 41 \tab 38 \tab 2 \tab 2 \tab 23 \tab 18 \tab 16 \tab 4 \tab 264 \tab 6 \tab 1 \tab - \tab 19 \tab 35 \tab 24 \tab 3 \tab 8 \tab - \tab 9 \tab - \tab 7 \tab 5 \tab 6 \tab 8 \tab 7 \tab 6 \tab 2 \tab - \tab - \tab 4 \tab 16 \tab 9 \tab - \tab 4 \tab 3 \tab 11 \tab 15 \tab 3 \tab 16 \tab 11 \tab 262 \tab 8 \tab 2 \tab 1 \tab 2 \tab - \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 2 \tab 2 \tab 9 \tab 1 \tab 3 \tab 14 \tab 30 \tab 5 \tab 57 \tab 23 \tab 10 \tab 19 \tab 3 \tab 24 \tab 17 \tab 4 \tab 22 \tab 20 \tab 16 \tab 55 \tab 5 \tab 5 \tab 13 \tab 8 \tab 22 \tab 18 \tab 24 \tab 4 \tab 15 \tab 8 \tab 2 \tab 53 \tab 10 \tab 6 \tab 15 \tab 4 \tab 8 \tab 43 \tab 29 \tab 10 \tab 5 \tab 11 \tab 17 \tab 4 \tab 3 \tab 258 \tab 2 \tab - \tab 4 \tab 9 \tab 33 \tab 11 \tab 5 \tab 9 \tab 4 \tab 2 \tab 3 \tab 1 \tab 5 \tab 14 \tab 13 \tab 4 \tab 10 \tab 4 \tab 1 \tab - \tab 7 \tab 16 \tab 8 \tab - \tab - \tab 8 \tab 4 \tab 11 \tab 3 \tab 9 \tab 7 \tab 365 \tab 11 \tab 1 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 3 \tab 4 \tab 13 \tab 3 \tab 2 \tab 15 \tab 26 \tab 3 \tab 62 \tab 26 \tab 14 \tab 14 \tab 1 \tab 16 \tab 16 \tab 9 \tab 28 \tab 15 \tab 33 \tab 64 \tab 14 \tab 4 \tab 6 \tab 9 \tab 20 \tab 16 \tab 36 \tab 9 \tab 37 \tab 8 \tab - \tab 55 \tab 6 \tab 10 \tab 13 \tab 3 \tab 7 \tab 69 \tab 29 \tab 11 \tab 5 \tab 15 \tab 22 \tab 26 \tab 3 \tab 231 \tab 6 \tab 1 \tab 1 \tab 12 \tab 40 \tab 22 \tab 4 \tab 7 \tab 5 \tab 3 \tab 1 \tab 5 \tab 7 \tab 7 \tab 12 \tab - \tab 10 \tab 6 \tab 2 \tab 1 \tab 9 \tab 25 \tab 6 \tab 1 \tab - \tab 10 \tab 10 \tab 11 \tab 3 \tab 15 \tab 18 \tab 220 \tab 11 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 2 \tab 8 \tab 8 \tab 3 \tab 1 \tab 13 \tab 22 \tab 3 \tab 55 \tab 17 \tab 7 \tab 11 \tab 9 \tab 18 \tab 12 \tab 8 \tab 19 \tab 17 \tab 20 \tab 65 \tab 3 \tab 5 \tab 9 \tab 7 \tab 15 \tab 17 \tab 24 \tab 13 \tab 47 \tab 8 \tab - \tab 46 \tab 11 \tab 5 \tab 8 \tab 2 \tab 3 \tab 45 \tab 43 \tab 7 \tab 2 \tab 15 \tab 23 \tab 16 \tab 5 \tab 216 \tab 4 \tab - \tab - \tab 13 \tab 32 \tab 20 \tab 3 \tab 15 \tab - \tab 8 \tab 1 \tab 7 \tab 7 \tab 11 \tab 13 \tab - \tab 11 \tab 3 \tab 1 \tab 3 \tab 8 \tab 13 \tab 5 \tab 2 \tab 1 \tab 12 \tab 7 \tab 8 \tab 4 \tab 20 \tab 15 \tab 282 \tab 12 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1 \tab 10 \tab 12 \tab 1 \tab 1 \tab 30 \tab 40 \tab 5 \tab 86 \tab 32 \tab 10 \tab 21 \tab 9 \tab 31 \tab 16 \tab 2 \tab 27 \tab 28 \tab 32 \tab 89 \tab 9 \tab 4 \tab 10 \tab 5 \tab 15 \tab 26 \tab 19 \tab 15 \tab 39 \tab 10 \tab 2 \tab 51 \tab 9 \tab 10 \tab 21 \tab 4 \tab 8 \tab 55 \tab 54 \tab 15 \tab 3 \tab 13 \tab 19 \tab 31 \tab 1 \tab 256 \tab 6 \tab - \tab 5 \tab 11 \tab 39 \tab 16 \tab 7 \tab 6 \tab 5 \tab 11 \tab 2 \tab 4 \tab 16 \tab 12 \tab 4 \tab 1 \tab 17 \tab 1 \tab 1 \tab - \tab 8 \tab 13 \tab 14 \tab 2 \tab - \tab 6 \tab 14 \tab 23 \tab 1 \tab 17 \tab 13 \tab 382 \tab 8 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1860 \cr
   1983 \tab 1 \tab 7 \tab 11 \tab 4 \tab 4 \tab 24 \tab 22 \tab 8 \tab 65 \tab 25 \tab 10 \tab 11 \tab 3 \tab 29 \tab 8 \tab 8 \tab 33 \tab 25 \tab 21 \tab 79 \tab 7 \tab 6 \tab 12 \tab 9 \tab 15 \tab 21 \tab 29 \tab 8 \tab 43 \tab 10 \tab 2 \tab 50 \tab 13 \tab 8 \tab 14 \tab 4 \tab 9 \tab 55 \tab 36 \tab 15 \tab 6 \tab 23 \tab 25 \tab 16 \tab - \tab 188 \tab 8 \tab 1 \tab - \tab 10 \tab 50 \tab 24 \tab 2 \tab 14 \tab 2 \tab 5 \tab 6 \tab 4 \tab 10 \tab 6 \tab 12 \tab 1 \tab 9 \tab 3 \tab 1 \tab 1 \tab 2 \tab 21 \tab 13 \tab 5 \tab 1 \tab 12 \tab 14 \tab 11 \tab 3 \tab 41 \tab 15 \tab 240 \tab 17 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1 \tab 6 \tab 10 \tab 2 \tab 1 \tab 14 \tab 23 \tab 2 \tab 52 \tab 20 \tab 6 \tab 13 \tab 3 \tab 22 \tab 10 \tab 5 \tab 25 \tab 25 \tab 30 \tab 72 \tab 3 \tab 6 \tab 8 \tab 12 \tab 17 \tab 18 \tab 17 \tab 11 \tab 48 \tab 11 \tab 1 \tab 43 \tab 1 \tab 11 \tab 7 \tab 8 \tab 7 \tab 33 \tab 40 \tab 10 \tab 4 \tab 12 \tab 24 \tab 19 \tab 4 \tab 232 \tab 5 \tab 1 \tab 3 \tab 14 \tab 41 \tab 26 \tab 4 \tab 8 \tab 3 \tab 8 \tab 2 \tab 5 \tab 11 \tab 13 \tab 6 \tab 1 \tab 9 \tab 8 \tab - \tab 2 \tab 6 \tab 21 \tab 9 \tab 6 \tab 2 \tab 6 \tab 15 \tab 13 \tab 4 \tab 20 \tab 6 \tab 221 \tab 17 \tab - \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1 \tab 4 \tab 7 \tab 3 \tab 4 \tab 14 \tab 21 \tab 5 \tab 61 \tab 22 \tab 6 \tab 12 \tab 5 \tab 33 \tab 3 \tab 3 \tab 17 \tab 12 \tab 31 \tab 77 \tab 11 \tab 4 \tab 10 \tab 7 \tab 17 \tab 14 \tab 18 \tab 8 \tab 48 \tab 11 \tab - \tab 36 \tab 4 \tab 8 \tab 10 \tab 3 \tab 7 \tab 54 \tab 37 \tab 14 \tab - \tab 12 \tab 20 \tab 14 \tab 4 \tab 229 \tab 6 \tab 1 \tab - \tab 14 \tab 47 \tab 25 \tab 1 \tab 16 \tab 3 \tab 10 \tab 5 \tab 5 \tab 10 \tab 6 \tab 10 \tab 1 \tab 12 \tab 4 \tab 3 \tab 1 \tab 11 \tab 20 \tab 15 \tab 4 \tab 1 \tab 10 \tab 15 \tab 19 \tab 4 \tab 23 \tab 15 \tab 248 \tab 16 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 8 \tab 13 \tab 4 \tab 6 \tab 18 \tab 20 \tab 3 \tab 74 \tab 10 \tab 9 \tab 15 \tab 4 \tab 20 \tab 11 \tab 5 \tab 19 \tab 21 \tab 34 \tab 56 \tab 9 \tab 3 \tab 9 \tab 8 \tab 18 \tab 8 \tab 31 \tab 7 \tab 42 \tab 8 \tab 1 \tab 37 \tab 9 \tab 6 \tab 16 \tab 2 \tab 7 \tab 44 \tab 37 \tab 9 \tab 1 \tab 16 \tab 23 \tab 27 \tab 7 \tab 200 \tab 3 \tab - \tab 3 \tab 11 \tab 41 \tab 21 \tab 2 \tab 12 \tab 4 \tab 8 \tab 1 \tab 9 \tab 11 \tab 6 \tab 10 \tab - \tab 12 \tab 2 \tab 3 \tab - \tab 6 \tab 14 \tab 11 \tab 3 \tab 2 \tab 8 \tab 13 \tab 12 \tab 2 \tab 25 \tab 7 \tab 253 \tab 13 \tab - \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 2 \tab 8 \tab 12 \tab 1 \tab 3 \tab 19 \tab 25 \tab 6 \tab 86 \tab 22 \tab 12 \tab 9 \tab 4 \tab 30 \tab 16 \tab 1 \tab 32 \tab 22 \tab 42 \tab 82 \tab 8 \tab 9 \tab 8 \tab 12 \tab 25 \tab 12 \tab 18 \tab 8 \tab 35 \tab 8 \tab - \tab 49 \tab 7 \tab 6 \tab 16 \tab 7 \tab 6 \tab 57 \tab 38 \tab 9 \tab - \tab 14 \tab 21 \tab 14 \tab - \tab 266 \tab 3 \tab - \tab - \tab 18 \tab 38 \tab 25 \tab 5 \tab 11 \tab 4 \tab 6 \tab 1 \tab 6 \tab 9 \tab 12 \tab 8 \tab 2 \tab 14 \tab 4 \tab 1 \tab - \tab 10 \tab 32 \tab 12 \tab 1 \tab 2 \tab 10 \tab 12 \tab 21 \tab 4 \tab 28 \tab 16 \tab 382 \tab 19 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab 4 \tab 2 \tab 2 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 2 \tab 11 \tab 7 \tab 2 \tab 2 \tab 17 \tab 22 \tab 3 \tab 65 \tab 19 \tab 7 \tab 11 \tab 2 \tab 20 \tab 8 \tab 3 \tab 17 \tab 28 \tab 39 \tab 61 \tab 9 \tab 2 \tab 11 \tab 5 \tab 12 \tab 8 \tab 28 \tab 8 \tab 62 \tab 9 \tab 1 \tab 32 \tab 11 \tab 6 \tab 13 \tab 4 \tab 4 \tab 43 \tab 38 \tab 13 \tab 6 \tab 12 \tab 22 \tab 23 \tab 1 \tab 197 \tab 3 \tab - \tab - \tab 8 \tab 37 \tab 27 \tab 7 \tab 11 \tab 4 \tab 6 \tab - \tab 6 \tab 11 \tab 9 \tab 10 \tab 2 \tab 10 \tab 2 \tab - \tab 1 \tab 7 \tab 23 \tab 11 \tab 1 \tab 4 \tab 6 \tab 20 \tab 16 \tab 8 \tab 26 \tab 12 \tab 227 \tab 23 \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 3 \tab 1 \tab 3 \tab - \tab - \tab 4 \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 3 \tab 5 \tab 18 \tab 2 \tab - \tab 8 \tab 21 \tab 3 \tab 60 \tab 15 \tab 5 \tab 11 \tab 6 \tab 27 \tab 6 \tab 6 \tab 22 \tab 17 \tab 27 \tab 91 \tab 11 \tab 8 \tab 7 \tab 4 \tab 10 \tab 15 \tab 27 \tab 6 \tab 33 \tab 9 \tab 2 \tab 42 \tab 10 \tab 7 \tab 13 \tab 7 \tab 3 \tab 33 \tab 47 \tab 12 \tab 5 \tab 15 \tab 24 \tab 18 \tab 3 \tab 210 \tab 7 \tab 1 \tab 1 \tab 11 \tab 48 \tab 24 \tab 4 \tab 13 \tab 2 \tab 16 \tab 2 \tab 3 \tab 10 \tab 11 \tab 9 \tab - \tab 17 \tab 1 \tab 3 \tab 2 \tab 8 \tab 18 \tab 12 \tab 1 \tab 1 \tab 6 \tab 18 \tab 24 \tab 5 \tab 27 \tab 17 \tab 247 \tab 19 \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 1 \tab 2 \tab - \tab 1 \tab 1 \tab - \tab 1537 \cr
   1990 \tab 2 \tab 5 \tab 8 \tab - \tab 3 \tab 11 \tab 23 \tab 4 \tab 45 \tab 18 \tab 9 \tab 6 \tab 2 \tab 24 \tab 4 \tab 3 \tab 29 \tab 20 \tab 34 \tab 62 \tab 15 \tab 4 \tab 17 \tab 8 \tab 24 \tab 7 \tab 21 \tab 6 \tab 45 \tab 12 \tab - \tab 32 \tab 6 \tab 7 \tab 13 \tab 12 \tab 7 \tab 43 \tab 41 \tab 8 \tab 6 \tab 10 \tab 11 \tab 19 \tab 4 \tab 180 \tab 6 \tab 1 \tab 1 \tab 11 \tab 34 \tab 11 \tab 3 \tab 12 \tab 6 \tab 5 \tab 1 \tab 6 \tab 8 \tab 10 \tab 4 \tab 1 \tab 8 \tab 4 \tab 1 \tab 2 \tab 6 \tab 16 \tab 9 \tab 3 \tab 3 \tab 2 \tab 15 \tab 10 \tab 3 \tab 30 \tab 10 \tab 233 \tab 13 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1 \tab 5 \tab 8 \tab 3 \tab 3 \tab 21 \tab 23 \tab 3 \tab 56 \tab 11 \tab 8 \tab 13 \tab 6 \tab 26 \tab 8 \tab 9 \tab 31 \tab 18 \tab 27 \tab 77 \tab 9 \tab 6 \tab 11 \tab 4 \tab 13 \tab 8 \tab 20 \tab 5 \tab 54 \tab 15 \tab 2 \tab 39 \tab 10 \tab 8 \tab 11 \tab 1 \tab 10 \tab 38 \tab 39 \tab 15 \tab 5 \tab 12 \tab 32 \tab 32 \tab 3 \tab 187 \tab 1 \tab - \tab - \tab 8 \tab 33 \tab 24 \tab 5 \tab 9 \tab 5 \tab 6 \tab 1 \tab 5 \tab 12 \tab 10 \tab 4 \tab 1 \tab 10 \tab 1 \tab 3 \tab 6 \tab 2 \tab 24 \tab 9 \tab 8 \tab 1 \tab 3 \tab 8 \tab 19 \tab 6 \tab 28 \tab 20 \tab 266 \tab 20 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab 1 \tab 1 \tab 3 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 5 \tab 6 \tab 2 \tab 2 \tab 20 \tab 32 \tab 5 \tab 74 \tab 21 \tab 6 \tab 17 \tab 5 \tab 15 \tab 8 \tab 6 \tab 32 \tab 18 \tab 32 \tab 71 \tab 14 \tab 8 \tab 15 \tab 8 \tab 18 \tab 14 \tab 27 \tab 13 \tab 45 \tab 12 \tab - \tab 35 \tab 8 \tab 10 \tab 9 \tab 2 \tab 5 \tab 43 \tab 38 \tab 8 \tab - \tab 14 \tab 16 \tab 25 \tab 3 \tab 205 \tab 5 \tab 1 \tab 2 \tab 18 \tab 38 \tab 28 \tab - \tab 13 \tab 2 \tab 8 \tab 2 \tab 17 \tab 17 \tab 11 \tab 16 \tab 1 \tab 22 \tab 3 \tab - \tab - \tab 5 \tab 20 \tab 12 \tab 4 \tab 3 \tab 12 \tab 18 \tab 23 \tab 7 \tab 25 \tab 22 \tab 244 \tab 20 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 3 \tab 4 \tab 2 \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 14 \tab 14 \tab 2 \tab 9 \tab 49 \tab 55 \tab 5 \tab 92 \tab 34 \tab 16 \tab 21 \tab 4 \tab 49 \tab 19 \tab 6 \tab 67 \tab 35 \tab 56 \tab 154 \tab 12 \tab 10 \tab 27 \tab 26 \tab 43 \tab 21 \tab 57 \tab 16 \tab 72 \tab 22 \tab 1 \tab 75 \tab 12 \tab 14 \tab 16 \tab 6 \tab 13 \tab 75 \tab 74 \tab 19 \tab 5 \tab 24 \tab 35 \tab 42 \tab 5 \tab 318 \tab 7 \tab 3 \tab 8 \tab 26 \tab 60 \tab 37 \tab 7 \tab 16 \tab 5 \tab 10 \tab 3 \tab 21 \tab 19 \tab 15 \tab 17 \tab 8 \tab 33 \tab 11 \tab 1 \tab 7 \tab 18 \tab 53 \tab 26 \tab 15 \tab 2 \tab 25 \tab 46 \tab 46 \tab 12 \tab 58 \tab 22 \tab 551 \tab 31 \tab 3 \tab 3 \tab 2 \tab 1 \tab 1 \tab 3 \tab 4 \tab 2 \tab 3 \tab 2 \tab 5 \tab - \tab 2 \tab 1 \tab - \tab 2992 \cr
   1996 \tab 3 \tab 12 \tab 22 \tab 5 \tab 4 \tab 34 \tab 47 \tab 2 \tab 100 \tab 34 \tab 17 \tab 21 \tab 7 \tab 57 \tab 16 \tab 12 \tab 50 \tab 20 \tab 64 \tab 128 \tab 12 \tab 14 \tab 15 \tab 16 \tab 36 \tab 18 \tab 60 \tab 15 \tab 78 \tab 22 \tab - \tab 71 \tab 9 \tab 14 \tab 14 \tab 6 \tab 9 \tab 65 \tab 77 \tab 18 \tab 7 \tab 18 \tab 44 \tab 49 \tab 8 \tab 270 \tab 10 \tab 1 \tab 2 \tab 24 \tab 78 \tab 55 \tab 7 \tab 30 \tab 5 \tab 15 \tab 3 \tab 15 \tab 22 \tab 9 \tab 19 \tab 2 \tab 34 \tab 7 \tab 6 \tab 8 \tab 8 \tab 55 \tab 23 \tab 10 \tab 7 \tab 18 \tab 41 \tab 48 \tab 10 \tab 59 \tab 25 \tab 568 \tab 34 \tab 1 \tab 4 \tab 1 \tab 3 \tab - \tab 2 \tab - \tab 5 \tab - \tab 2 \tab 4 \tab - \tab 3 \tab 1 \tab - \tab 2904 \cr
   1998 \tab 2 \tab 8 \tab 15 \tab 3 \tab 6 \tab 32 \tab 30 \tab 3 \tab 88 \tab 36 \tab 16 \tab 18 \tab 8 \tab 35 \tab 20 \tab 8 \tab 57 \tab 43 \tab 58 \tab 113 \tab 16 \tab 7 \tab 22 \tab 10 \tab 25 \tab 29 \tab 60 \tab 20 \tab 80 \tab 26 \tab - \tab 74 \tab 11 \tab 15 \tab 15 \tab 4 \tab 14 \tab 66 \tab 70 \tab 24 \tab 11 \tab 19 \tab 48 \tab 30 \tab 6 \tab 255 \tab 5 \tab 4 \tab 6 \tab 32 \tab 74 \tab 45 \tab 7 \tab 29 \tab 11 \tab 15 \tab 3 \tab 13 \tab 28 \tab 12 \tab 25 \tab - \tab 24 \tab 7 \tab 4 \tab - \tab 9 \tab 48 \tab 22 \tab 11 \tab 5 \tab 22 \tab 32 \tab 47 \tab 9 \tab 65 \tab 28 \tab 569 \tab 33 \tab 1 \tab 1 \tab - \tab 3 \tab - \tab 3 \tab 5 \tab 5 \tab 2 \tab 1 \tab 8 \tab 1 \tab 1 \tab 1 \tab - \tab 2832 \cr
   2000 \tab 3 \tab 16 \tab 16 \tab 3 \tab 5 \tab 37 \tab 46 \tab 6 \tab 92 \tab 38 \tab 25 \tab 25 \tab 10 \tab 49 \tab 13 \tab 13 \tab 37 \tab 35 \tab 67 \tab 151 \tab 10 \tab 6 \tab 18 \tab 7 \tab 31 \tab 19 \tab 38 \tab 16 \tab 68 \tab 16 \tab 4 \tab 65 \tab 16 \tab 20 \tab 16 \tab 8 \tab 7 \tab 64 \tab 81 \tab 13 \tab 8 \tab 24 \tab 33 \tab 43 \tab 5 \tab 274 \tab 16 \tab 3 \tab 1 \tab 24 \tab 61 \tab 35 \tab 7 \tab 19 \tab 6 \tab 11 \tab 8 \tab 8 \tab 41 \tab 16 \tab 12 \tab 3 \tab 35 \tab 12 \tab 1 \tab 2 \tab 9 \tab 38 \tab 21 \tab 18 \tab 2 \tab 17 \tab 29 \tab 42 \tab 7 \tab 60 \tab 30 \tab 576 \tab 27 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab 3 \tab 1 \tab 6 \tab - \tab 1 \tab 4 \tab - \tab 1 \tab - \tab - \tab 2817 \cr
   2002 \tab 1 \tab 11 \tab 9 \tab 1 \tab 6 \tab 28 \tab 42 \tab 4 \tab 91 \tab 34 \tab 16 \tab 22 \tab 7 \tab 46 \tab 17 \tab 11 \tab 38 \tab 32 \tab 79 \tab 132 \tab 16 \tab 9 \tab 21 \tab 8 \tab 30 \tab 17 \tab 38 \tab 12 \tab 67 \tab 24 \tab 2 \tab 64 \tab 10 \tab 13 \tab 17 \tab 10 \tab 10 \tab 67 \tab 75 \tab 23 \tab 12 \tab 17 \tab 50 \tab 35 \tab 3 \tab 234 \tab 14 \tab 2 \tab 7 \tab 13 \tab 64 \tab 40 \tab 4 \tab 26 \tab 4 \tab 11 \tab 7 \tab 16 \tab 25 \tab 26 \tab 27 \tab 4 \tab 28 \tab 9 \tab 1 \tab 5 \tab 11 \tab 52 \tab 17 \tab 11 \tab 3 \tab 13 \tab 28 \tab 48 \tab 8 \tab 49 \tab 24 \tab 589 \tab 36 \tab 1 \tab 6 \tab 2 \tab 4 \tab 2 \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 8 \tab - \tab 1 \tab 1 \tab - \tab 2765 \cr
   2004 \tab 1 \tab 12 \tab 9 \tab - \tab 1 \tab 32 \tab 34 \tab 4 \tab 117 \tab 37 \tab 16 \tab 17 \tab 6 \tab 64 \tab 10 \tab 11 \tab 52 \tab 40 \tab 71 \tab 127 \tab 12 \tab 14 \tab 23 \tab 10 \tab 29 \tab 19 \tab 51 \tab 19 \tab 64 \tab 13 \tab - \tab 71 \tab 11 \tab 12 \tab 17 \tab 11 \tab 5 \tab 70 \tab 79 \tab 20 \tab 8 \tab 25 \tab 36 \tab 30 \tab 8 \tab 197 \tab 10 \tab 1 \tab 4 \tab 19 \tab 76 \tab 30 \tab 8 \tab 22 \tab 5 \tab 8 \tab 6 \tab 16 \tab 29 \tab 24 \tab 19 \tab 3 \tab 27 \tab 10 \tab 3 \tab 7 \tab 4 \tab 55 \tab 18 \tab 11 \tab 7 \tab 22 \tab 28 \tab 47 \tab 12 \tab 63 \tab 35 \tab 603 \tab 30 \tab 1 \tab 4 \tab - \tab 2 \tab 1 \tab 4 \tab 6 \tab 6 \tab - \tab - \tab 5 \tab 1 \tab 5 \tab - \tab - \tab 2812 \cr
   2006 \tab 3 \tab 11 \tab 16 \tab 2 \tab 9 \tab 35 \tab 40 \tab 4 \tab 118 \tab 50 \tab 14 \tab 32 \tab 13 \tab 63 \tab 15 \tab 13 \tab 57 \tab 39 \tab 72 \tab 133 \tab 11 \tab 6 \tab 25 \tab 14 \tab 34 \tab 23 \tab 39 \tab 14 \tab 73 \tab 19 \tab 2 \tab 69 \tab 5 \tab 16 \tab 21 \tab 6 \tab 11 \tab 63 \tab 86 \tab 21 \tab 12 \tab 29 \tab 56 \tab 39 \tab 5 \tab 218 \tab 10 \tab 1 \tab 10 \tab 34 \tab 52 \tab 54 \tab 9 \tab 33 \tab 10 \tab 8 \tab 5 \tab 7 \tab 27 \tab 35 \tab 16 \tab 2 \tab 30 \tab 5 \tab 2 \tab 3 \tab 16 \tab 62 \tab 28 \tab 10 \tab 6 \tab 23 \tab 35 \tab 39 \tab 12 \tab 56 \tab 38 \tab 2107 \tab 40 \tab - \tab 5 \tab 2 \tab 1 \tab - \tab 3 \tab 5 \tab 3 \tab 1 \tab 3 \tab 4 \tab - \tab 1 \tab 1 \tab - \tab 4510 \cr
   2008 \tab - \tab 11 \tab 13 \tab - \tab 2 \tab 27 \tab 26 \tab 7 \tab 57 \tab 28 \tab 11 \tab 12 \tab 3 \tab 41 \tab 10 \tab 8 \tab 28 \tab 28 \tab 63 \tab 84 \tab 14 \tab 6 \tab 16 \tab 17 \tab 19 \tab 16 \tab 43 \tab 11 \tab 29 \tab 15 \tab 1 \tab 58 \tab 8 \tab 9 \tab 8 \tab 4 \tab 8 \tab 40 \tab 66 \tab 14 \tab 7 \tab 9 \tab 42 \tab 20 \tab 6 \tab 144 \tab 6 \tab 3 \tab 4 \tab 29 \tab 42 \tab 32 \tab 12 \tab 15 \tab 9 \tab 14 \tab 3 \tab 13 \tab 28 \tab 29 \tab 13 \tab 3 \tab 20 \tab 3 \tab 6 \tab 4 \tab 6 \tab 40 \tab 11 \tab 4 \tab 6 \tab 11 \tab 22 \tab 23 \tab 5 \tab 50 \tab 19 \tab 398 \tab 30 \tab 1 \tab 3 \tab - \tab 1 \tab - \tab 6 \tab 2 \tab 2 \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 1 \tab - \tab 2023 \cr
   2010 \tab - \tab 13 \tab 12 \tab 2 \tab 1 \tab 29 \tab 30 \tab 6 \tab 68 \tab 19 \tab 12 \tab 19 \tab 10 \tab 51 \tab 12 \tab 5 \tab 29 \tab 27 \tab 35 \tab 98 \tab 8 \tab 8 \tab 18 \tab 5 \tab 16 \tab 15 \tab 27 \tab 17 \tab 52 \tab 20 \tab 1 \tab 57 \tab 9 \tab 15 \tab 17 \tab 7 \tab 8 \tab 44 \tab 64 \tab 12 \tab 5 \tab 26 \tab 22 \tab 28 \tab 11 \tab 135 \tab 8 \tab 2 \tab 3 \tab 13 \tab 47 \tab 28 \tab 8 \tab 24 \tab 4 \tab 7 \tab 8 \tab 9 \tab 16 \tab 19 \tab 15 \tab 2 \tab 17 \tab 7 \tab 3 \tab 4 \tab 5 \tab 37 \tab 8 \tab 11 \tab 6 \tab 11 \tab 23 \tab 29 \tab 4 \tab 55 \tab 28 \tab 402 \tab 33 \tab 1 \tab 4 \tab - \tab - \tab 1 \tab - \tab 5 \tab 1 \tab - \tab 2 \tab 7 \tab - \tab 1 \tab 1 \tab - \tab 2044 \cr
   2012 \tab 3 \tab 4 \tab 11 \tab 1 \tab 3 \tab 31 \tab 30 \tab 3 \tab 77 \tab 20 \tab 13 \tab 9 \tab 2 \tab 60 \tab 15 \tab 7 \tab 31 \tab 25 \tab 66 \tab 81 \tab 8 \tab 6 \tab 12 \tab 3 \tab 26 \tab 11 \tab 30 \tab 5 \tab 47 \tab 9 \tab 1 \tab 64 \tab 9 \tab 8 \tab 2 \tab 8 \tab 5 \tab 41 \tab 53 \tab 16 \tab 1 \tab 22 \tab 24 \tab 26 \tab 2 \tab 123 \tab 9 \tab 1 \tab 5 \tab 19 \tab 29 \tab 20 \tab 8 \tab 16 \tab 5 \tab 5 \tab 2 \tab 11 \tab 24 \tab - \tab 15 \tab 5 \tab 27 \tab 4 \tab 5 \tab 2 \tab 10 \tab 73 \tab 16 \tab 11 \tab 5 \tab 10 \tab 19 \tab 30 \tab 7 \tab 50 \tab 49 \tab 385 \tab 16 \tab 2 \tab 6 \tab 1 \tab 1 \tab - \tab 3 \tab 4 \tab 1 \tab - \tab - \tab 8 \tab - \tab - \tab 1 \tab - \tab 1974 \cr
   2014 \tab 1 \tab 13 \tab 10 \tab 3 \tab - \tab 33 \tab 44 \tab 2 \tab 97 \tab 24 \tab 11 \tab 9 \tab 5 \tab 61 \tab 16 \tab 4 \tab 51 \tab 32 \tab 67 \tab 112 \tab 14 \tab 7 \tab 19 \tab 10 \tab 26 \tab 16 \tab 30 \tab 13 \tab 71 \tab 22 \tab 1 \tab 47 \tab 25 \tab 18 \tab 9 \tab 6 \tab 5 \tab 62 \tab 83 \tab 19 \tab 10 \tab 17 \tab 37 \tab 48 \tab 5 \tab 156 \tab 11 \tab 1 \tab 3 \tab 13 \tab 60 \tab 34 \tab 8 \tab 30 \tab 7 \tab 6 \tab 6 \tab 7 \tab 32 \tab 50 \tab 24 \tab 1 \tab 19 \tab 10 \tab 4 \tab 3 \tab 5 \tab 37 \tab 18 \tab 15 \tab 4 \tab 19 \tab 29 \tab 42 \tab 5 \tab 56 \tab 23 \tab 524 \tab 32 \tab - \tab 5 \tab 3 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 8 \tab - \tab 3 \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 16 \tab 17 \tab 2 \tab 1 \tab 58 \tab 36 \tab 4 \tab 121 \tab 23 \tab 18 \tab 22 \tab 9 \tab 82 \tab 19 \tab 16 \tab 48 \tab 36 \tab 75 \tab 138 \tab 12 \tab 8 \tab 24 \tab 13 \tab 36 \tab 9 \tab 45 \tab 13 \tab 59 \tab 20 \tab 5 \tab 71 \tab 21 \tab 14 \tab 13 \tab 6 \tab 8 \tab 58 \tab 95 \tab 12 \tab 12 \tab 14 \tab 38 \tab 22 \tab 8 \tab 144 \tab 17 \tab 1 \tab 7 \tab 17 \tab 55 \tab 30 \tab 12 \tab 42 \tab 9 \tab 11 \tab 12 \tab 18 \tab 25 \tab 26 \tab 34 \tab 2 \tab 21 \tab 13 \tab 1 \tab 7 \tab 24 \tab 42 \tab 16 \tab 14 \tab 9 \tab 15 \tab 30 \tab 53 \tab 4 \tab 63 \tab 28 \tab 625 \tab 28 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab 4 \tab 5 \tab 7 \tab 1 \tab 3 \tab 7 \tab - \tab 1 \tab 1 \tab - \tab 2867 \cr
   2018 \tab - \tab 8 \tab 12 \tab 1 \tab 7 \tab 36 \tab 37 \tab 14 \tab 90 \tab 22 \tab 14 \tab 20 \tab 10 \tab 56 \tab 14 \tab 7 \tab 34 \tab 25 \tab 54 \tab 106 \tab 12 \tab 13 \tab 18 \tab 6 \tab 29 \tab 8 \tab 26 \tab 14 \tab 39 \tab 8 \tab 1 \tab 39 \tab 15 \tab 14 \tab 10 \tab 6 \tab 8 \tab 49 \tab 70 \tab 21 \tab 43 \tab 22 \tab 36 \tab 26 \tab 10 \tab 114 \tab 14 \tab 4 \tab 8 \tab 20 \tab 37 \tab 63 \tab 4 \tab 28 \tab 5 \tab 14 \tab 6 \tab 10 \tab 26 \tab 28 \tab 18 \tab 1 \tab 32 \tab 9 \tab 4 \tab 6 \tab 12 \tab 31 \tab 16 \tab 13 \tab 8 \tab 11 \tab 39 \tab 32 \tab 3 \tab 35 \tab 34 \tab 498 \tab 23 \tab 2 \tab 4 \tab - \tab 1 \tab 2 \tab 4 \tab 7 \tab 2 \tab 3 \tab - \tab 4 \tab - \tab 3 \tab - \tab - \tab 2348 \cr
   2021 \tab 1 \tab 19 \tab 20 \tab 3 \tab 3 \tab 49 \tab 60 \tab 3 \tab 114 \tab 39 \tab 22 \tab 20 \tab 6 \tab 85 \tab 36 \tab 22 \tab 79 \tab 58 \tab 107 \tab 206 \tab 14 \tab 18 \tab 37 \tab 18 \tab 42 \tab 24 \tab 77 \tab 24 \tab 112 \tab 29 \tab 1 \tab 147 \tab 24 \tab 21 \tab 9 \tab 11 \tab 24 \tab 100 \tab 128 \tab 41 \tab 6 \tab 38 \tab 67 \tab 57 \tab 15 \tab 167 \tab 22 \tab 5 \tab 9 \tab 27 \tab 80 \tab 44 \tab 11 \tab 44 \tab 19 \tab 17 \tab 11 \tab 24 \tab 34 \tab 88 \tab 51 \tab 3 \tab 34 \tab 15 \tab 6 \tab 9 \tab 15 \tab 81 \tab 34 \tab 22 \tab 13 \tab 37 \tab 58 \tab 72 \tab 14 \tab 110 \tab 34 \tab 667 \tab 50 \tab 1 \tab 13 \tab 6 \tab 2 \tab - \tab 7 \tab 8 \tab 4 \tab - \tab - \tab 24 \tab - \tab 3 \tab 1 \tab - \tab 4032 \cr
   2022 \tab 1 \tab 13 \tab 9 \tab 2 \tab 7 \tab 43 \tab 75 \tab 5 \tab 121 \tab 25 \tab 23 \tab 14 \tab 6 \tab 104 \tab 17 \tab 16 \tab 65 \tab 40 \tab 85 \tab 177 \tab 12 \tab 10 \tab 40 \tab 21 \tab 23 \tab 21 \tab 44 \tab 19 \tab 76 \tab 22 \tab 3 \tab 104 \tab 22 \tab 13 \tab 16 \tab 8 \tab 14 \tab 80 \tab 100 \tab 24 \tab 6 \tab 20 \tab 51 \tab 49 \tab 11 \tab 140 \tab 19 \tab - \tab 5 \tab 25 \tab 56 \tab 40 \tab 14 \tab 41 \tab 9 \tab 16 \tab 4 \tab 11 \tab 48 \tab 83 \tab 54 \tab 1 \tab 43 \tab 16 \tab - \tab 18 \tab 9 \tab 63 \tab 25 \tab 18 \tab 8 \tab 28 \tab 44 \tab 58 \tab 11 \tab 73 \tab 37 \tab 770 \tab 44 \tab 3 \tab 8 \tab 3 \tab 3 \tab - \tab 9 \tab 7 \tab 3 \tab 3 \tab 2 \tab 11 \tab - \tab 3 \tab 1 \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 47 \tab 303 \tab 379 \tab 73 \tab 115 \tab 864 \tab 1108 \tab 150 \tab 2693 \tab 854 \tab 400 \tab 540 \tab 193 \tab 1349 \tab 457 \tab 265 \tab 1199 \tab 894 \tab 1553 \tab 3259 \tab 336 \tab 243 \tab 527 \tab 329 \tab 738 \tab 532 \tab 1104 \tab 406 \tab 1734 \tab 472 \tab 40 \tab 1890 \tab 348 \tab 364 \tab 446 \tab 195 \tab 267 \tab 1848 \tab 1952 \tab 480 \tab 212 \tab 608 \tab 996 \tab 910 \tab 164 \tab 7260 \tab 257 \tab 46 \tab 119 \tab 567 \tab 1582 \tab 986 \tab 190 \tab 619 \tab 173 \tab 308 \tab 123 \tab 303 \tab 598 \tab 631 \tab 535 \tab 75 \tab 635 \tab 201 \tab 73 \tab 107 \tab 282 \tab 1123 \tab 492 \tab 245 \tab 124 \tab 408 \tab 740 \tab 934 \tab 195 \tab 1325 \tab 690 \tab 15147 \tab 749 \tab 32 \tab 102 \tab 41 \tab 37 \tab 12 \tab 89 \tab 88 \tab 72 \tab 24 \tab 26 \tab 140 \tab 5 \tab 32 \tab 12 \tab 3309 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{maind80}},
\code{\link{maocc10}},
\code{\link{maocc80}},
\code{\link{mapres80}},
\code{\link{mawrkslf}},
\code{\link{paind10}},
\code{\link{paindstatus}},
\code{\link{paindtag}},
\code{\link{papres10}}
}
\concept{Core}
\keyword{variable}
