% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_07.R
\name{papres80}
\alias{papres80}
\title{Father's occupational prestige score (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
papres80
}
\details{
Question Father's Occupation, Prestige, and Industry
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/41/vshow}{GSS Data Explorer page for \code{papres80}}.
}

\seealso{
Other Core: 
\code{\link{paind16}},
\code{\link{paind80}},
\code{\link{paocc10}},
\code{\link{paocc80}},
\code{\link{paoccindv}},
\code{\link{paoccstatus}},
\code{\link{paocctag}},
\code{\link{papres16}},
\code{\link{pawrkslf}}
}
\concept{Core}
\keyword{variable}
