% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_534.R
\name{parcit}
\alias{parcit}
\title{Were your parents citizens of america?}
\source{
General Social Survey https://gss.norc.org
}
\description{
parcit
}
\details{
Question B. At the time of your birth, were both, one or neither of your parents citizens of America?
}
\section{Values}{

\itemize{
\item \verb{[1]} both were citizens of america
\item \verb{[2]} only father was a citizen of america
\item \verb{[3]} only mother was a citizen of america
\item \verb{[4]} neither was a citizen of america
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4874/vshow}{GSS Data Explorer page for \code{parcit}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab both were citizens of america \tab only father was a citizen of america \tab only mother was a citizen of america \tab neither was a citizen of america \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 1218 \tab 17 \tab 13 \tab 99 \tab 2 \tab 18 \tab 2904 \cr
   2004 \tab 1596 \tab 1065 \tab 18 \tab 17 \tab 113 \tab 2 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 1048 \tab 29 \tab 13 \tab 174 \tab 6 \tab 4 \tab 2538 \cr
   Total \tab 4397 \tab 3331 \tab 64 \tab 43 \tab 386 \tab 10 \tab 23 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{citizen}},
\code{\link{ethclose}},
\code{\link{nafta1}},
\code{\link{nafta2}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang1}},
\code{\link{spklang2}},
\code{\link{spklang3}}
}
\concept{ISSP National Identity}
\keyword{variable}
