% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_115.R
\name{parelig}
\alias{parelig}
\title{What was father's religious preference}
\source{
General Social Survey https://gss.norc.org
}
\description{
parelig
}
\details{
Question What was your father''s religious preference when you were growing up?  Was it Protestant, Catholic, Jewish, some other religion, or no religion?
}
\section{Values}{

\itemize{
\item \code{1} protestant
\item \code{2} catholic
\item \code{3} jewish
\item \code{4} none
\item \code{5} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1063/vshow}{GSS Data Explorer page for \code{parelig}}.
}

\seealso{
Other Religion: 
\code{\link{churhsch}},
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{maoth}},
\code{\link{paden}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{paoth}},
\code{\link{sunsch16}}
}
\concept{Religion}
\keyword{variable}
