% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_557.R
\name{parfin}
\alias{parfin}
\title{Hard for parents to make ends meet}
\source{
General Social Survey https://gss.norc.org
}
\description{
parfin
}
\details{
Question Please tell me whether you strongly agree, agree, disagree, or strongly disagree with the following statements regarding parents:
E. Parents today have a hard time making ends meet.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5097/vshow}{GSS Data Explorer page for \code{parfin}}.
}

\seealso{
Other Children: 
\code{\link{chldeduc}},
\code{\link{chldhlth}},
\code{\link{chldhome}},
\code{\link{chldlove}},
\code{\link{chldmorl}},
\code{\link{chldsafe}},
\code{\link{chldskls}},
\code{\link{chldsup}},
\code{\link{chldtime}},
\code{\link{infadstv}},
\code{\link{inffilms}},
\code{\link{infmusic}},
\code{\link{infnettv}},
\code{\link{infpubtv}},
\code{\link{pargovt}},
\code{\link{parhardr}},
\code{\link{parrght}},
\code{\link{partaxes}},
\code{\link{partime}},
\code{\link{parwhere}},
\code{\link{parwork}},
\code{\link{parworse}},
\code{\link{roleccp}},
\code{\link{roleclrg}},
\code{\link{rolegp}},
\code{\link{rolema}},
\code{\link{rolepa}},
\code{\link{roletchr}},
\code{\link{spdrugs}},
\code{\link{spdsabkd}},
\code{\link{spfoodkd}},
\code{\link{spheadst}},
\code{\link{sphlthkd}},
\code{\link{sphomekd}},
\code{\link{sppill}},
\code{\link{sppoorkd}},
\code{\link{sppregnt}},
\code{\link{spwrkpar}}
}
\concept{Children}
\keyword{variable}
