% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_179.R
\name{parleave}
\alias{parleave}
\title{Offer paid time off to new parents.}
\source{
General Social Survey https://gss.norc.org
}
\description{
parleave
}
\details{
Question It is becoming more common for employers to offer ways for people to combine work and being a parent. For each of the following, please tell me whether or not you think that employers should offer these options to their workers.
B. Offer paid time off to new parents.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, should offer this option
\item \verb{[2]} no, should not offer this option
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1677/vshow}{GSS Data Explorer page for \code{parleave}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes, should offer this option \tab no, should not offer this option \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 1057 \tab 338 \tab 60 \tab 5 \tab 2904 \cr
   Total \tab 1444 \tab 1057 \tab 338 \tab 60 \tab 5 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{allben}},
\code{\link{chldben}},
\code{\link{discaffm}},
\code{\link{discaffw}},
\code{\link{flextime}},
\code{\link{menben}},
\code{\link{nohmewrk}},
\code{\link{nonurse}},
\code{\link{womenben}}
}
\concept{Gender}
\keyword{variable}
