% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_557.R
\name{parrght}
\alias{parrght}
\title{Parents unsure what is right}
\source{
General Social Survey https://gss.norc.org
}
\description{
parrght
}
\details{
Question Please tell me whether you strongly agree, agree, disagree, or strongly disagree with the following statements regarding parents:
C. Parents today often feel uncertain about what is the right thing to do in raising their children.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5095/vshow}{GSS Data Explorer page for \code{parrght}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 245 \tab 926 \tab 122 \tab 27 \tab 35 \tab 17 \tab 1372 \cr
   Total \tab 0 \tab 245 \tab 926 \tab 122 \tab 27 \tab 35 \tab 17 \tab 1372 \cr
}
}

\seealso{
Other Children: 
\code{\link{parfin}},
\code{\link{pargovt}},
\code{\link{parhardr}},
\code{\link{partaxes}},
\code{\link{partime}},
\code{\link{parwhere}},
\code{\link{parwork}},
\code{\link{parworse}},
\code{\link{rolema}}
}
\concept{Children}
\keyword{variable}
