% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_65.R
\name{parsol}
\alias{parsol}
\title{R's living standard compared to parents}
\source{
General Social Survey https://gss.norc.org
}
\description{
parsol
}
\details{
Question A. Compared to your parents when they were the age you are now, Do you think your own standard of living now is much better, somewhat better, about the same, somewhat worse, or much worse than theirs was?
}
\section{Values}{

\itemize{
\item \code{1} much better
\item \code{2} somewhat better
\item \code{3} about the same
\item \code{4} somewhat worse
\item \code{5} much worse
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab B/C \tab partial \cr
   1996 \tab B/C \tab partial \cr
   1998 \tab B/C \tab partial \cr
   2000 \tab B/C \tab partial \cr
   2002 \tab B/C \tab partial \cr
   2004 \tab B/C \tab partial \cr
   2006 \tab B/C \tab partial \cr
   2008 \tab B/C \tab partial \cr
   2010 \tab B/C \tab partial \cr
   2012 \tab B/C \tab partial \cr
   2014 \tab B/C \tab partial \cr
   2016 \tab B/C \tab partial \cr
   2018 \tab B/C \tab partial \cr
   2021 \tab B/C \tab partial \cr
   2022 \tab B/C \tab partial \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/586/vshow}{GSS Data Explorer page for \code{parsol}}.
}

\seealso{
Other Core: 
\code{\link{getahead}},
\code{\link{getaid}},
\code{\link{govaid}},
\code{\link{unemp}},
\code{\link{union}()},
\code{\link{union1}},
\code{\link{wksup1}},
\code{\link{wksups}},
\code{\link{wksups1}}
}
\concept{Core}
\concept{Parents}
\concept{Social Mobility}
\keyword{variable}
