% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_59.R
\name{partfull}
\alias{partfull}
\title{Was r's work part-time or full-time?}
\source{
General Social Survey https://gss.norc.org
}
\description{
partfull
}
\details{
Question Now I'd like to ask you about last year. In (the previous year) how many weeks did you work either full-time or part-time not counting work around the house--include paid vacations and sick leave?
A. When you worked in (the previous year), was it usually full or part time?
}
\section{Values}{

\itemize{
\item \code{1} full-time
\item \code{2} part-time
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/529/vshow}{GSS Data Explorer page for \code{partfull}}.
}

\seealso{
Other Core: 
\code{\link{aged}},
\code{\link{agedchld}},
\code{\link{agedpar}},
\code{\link{drink}},
\code{\link{drunk}},
\code{\link{smoke}},
\code{\link{socpars}},
\code{\link{socsibs}},
\code{\link{weekswrk}}

Other Employment: 
\code{\link{weekswrk}}

Other Labor Force Status: 
\code{\link{weekswrk}}
}
\concept{Core}
\concept{Employment}
\concept{Labor Force Status}
\keyword{variable}
