% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_59.R
\name{partfull}
\alias{partfull}
\title{Was r's work part-time or full-time?}
\source{
General Social Survey https://gss.norc.org
}
\description{
partfull
}
\details{
Question Now I'd like to ask you about last year. In (the previous year) how many weeks did you work either full-time or part-time not counting work around the house--include paid vacations and sick leave?
A. When you worked in (the previous year), was it usually full or part time?
}
\section{Values}{

\itemize{
\item \code{1} full-time
\item \code{2} part-time
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/529/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab don't know \tab full-time \tab no answer \tab part-time \tab skipped on web \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 806 \tab 2 \tab 1717 \tab 52 \tab 415 \tab - \tab - \tab 2992 \cr
   1996 \tab 699 \tab 2 \tab 1756 \tab 57 \tab 390 \tab - \tab - \tab 2904 \cr
   1998 \tab 750 \tab 5 \tab 1638 \tab 70 \tab 369 \tab - \tab - \tab 2832 \cr
   2000 \tab 734 \tab 1 \tab 1645 \tab 49 \tab 388 \tab - \tab - \tab 2817 \cr
   2002 \tab 712 \tab 1 \tab 1619 \tab 6 \tab 427 \tab - \tab - \tab 2765 \cr
   2004 \tab 718 \tab - \tab 1633 \tab 4 \tab 457 \tab - \tab - \tab 2812 \cr
   2006 \tab 2420 \tab - \tab 1669 \tab - \tab 421 \tab - \tab - \tab 4510 \cr
   2008 \tab 564 \tab 2 \tab 1142 \tab 1 \tab 314 \tab - \tab - \tab 2023 \cr
   2010 \tab 634 \tab 1 \tab 1064 \tab 7 \tab 338 \tab - \tab - \tab 2044 \cr
   2012 \tab 621 \tab 1 \tab 1021 \tab - \tab 331 \tab - \tab - \tab 1974 \cr
   2014 \tab 794 \tab 3 \tab 1359 \tab - \tab 382 \tab - \tab - \tab 2538 \cr
   2016 \tab 949 \tab 1 \tab 1471 \tab 1 \tab 445 \tab - \tab - \tab 2867 \cr
   2018 \tab 749 \tab 1 \tab 1257 \tab 1 \tab 340 \tab - \tab - \tab 2348 \cr
   2021 \tab 1280 \tab 2 \tab 2145 \tab - \tab 602 \tab 3 \tab - \tab 4032 \cr
   2022 \tab 1227 \tab 2 \tab 1817 \tab - \tab 496 \tab 2 \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 43045 \tab 24 \tab 22953 \tab 248 \tab 6115 \tab 5 \tab 3309 \tab 75699 \cr
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\seealso{
Other Core: 
\code{\link{aged}},
\code{\link{agedchld}},
\code{\link{agedpar}},
\code{\link{drink}},
\code{\link{drunk}},
\code{\link{smoke}},
\code{\link{socpars}},
\code{\link{socsibs}},
\code{\link{weekswrk}}

Other Employment: 
\code{\link{weekswrk}}
}
\concept{Core}
\concept{Employment}
\concept{Labor Force Status}
\keyword{variable}
