% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_552.R
\name{partners}
\alias{partners}
\title{How many sex partner's r had in last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
partners
}
\details{
Question How many sex partners have you had in the last 12 months?
As of 2024, PARTNERS has been recoded to only account for 0 to 11+ partners and has been slightly adjusted to help protect confidentiality. Please see the 2024 codebook for more information.The original PARTNERS containing data from 1988 to 2022 has been renamed PARTNERS_8822.
}
\section{Values}{

\itemize{
\item \code{0} no partners
\item \code{1} 1 partner
\item \code{2} 2 partners
\item \code{3} 3 partners
\item \code{4} 4 partners
\item \code{5} 5-10 partners
\item \code{6} 11-20 partners
\item \code{7} 21-100 partners
\item \code{8} more than 100 partners
\item \code{9} 1 or more (unspecified)
\item \code{95} several
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab B/C/- \tab partial \cr
   2018 \tab B/C/- \tab partial \cr
   2021 \tab B/C/- \tab partial \cr
   2022 \tab B/C/- \tab partial \cr
   2024 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5049/vshow}{GSS Data Explorer page for \code{partners}}.
}

\seealso{
Other High Risk Behaviors: 
\code{\link{acqntsex}},
\code{\link{frndsex}},
\code{\link{matesex}},
\code{\link{othersex}},
\code{\link{paidsex}},
\code{\link{pikupsex}},
\code{\link{sexfreq}},
\code{\link{sexsex}}

Other Sex: 
\code{\link{acqntsex}},
\code{\link{frndsex}},
\code{\link{matesex}},
\code{\link{othersex}},
\code{\link{paidsex}},
\code{\link{pikupsex}},
\code{\link{sexfreq}},
\code{\link{sexsex}}

Other Split Ballots: 
\code{\link{aidsfare}},
\code{\link{sexfreq}}
}
\concept{High Risk Behaviors}
\concept{Sex}
\concept{Split Ballots}
\keyword{variable}
