% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_553.R
\name{partnrs5}
\alias{partnrs5}
\title{How many sex partner's r had in last 5 years}
\source{
General Social Survey https://gss.norc.org
}
\description{
partnrs5
}
\details{
Question Now think about the past five years--the time since February/March (1999), and including the past 12 months, how many sex partners have you had in that five year period?
}
\section{Values}{

\itemize{
\item \code{0} no partners
\item \code{1} 1 partner
\item \code{2} 2 partners
\item \code{3} 3 partners
\item \code{4} 4 partners
\item \code{5} 5-10 partners
\item \code{6} 11-20 partners
\item \code{7} 21-100 partners
\item \code{8} more than 100 partners
\item \code{9} 1 or more (unspecified)
\item \code{95} several
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab B/C/- \tab partial \cr
   2018 \tab B/C/- \tab partial \cr
   2021 \tab B/C/- \tab partial \cr
   2022 \tab B/C/- \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5064/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab 1 or more (unspecified) \tab 1 partner \tab 11-20 partners \tab 2 partners \tab 21-100 partners \tab 3 partners \tab 4 partners \tab 5-10 partners \tab more than 100 partners \tab no answer \tab no partners \tab don't know \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 872 \tab 1 \tab 348 \tab 11 \tab 51 \tab 10 \tab 48 \tab 28 \tab 37 \tab 2 \tab 14 \tab 95 \tab - \tab - \tab 1517 \cr
   1993 \tab 114 \tab 4 \tab 793 \tab 34 \tab 136 \tab 13 \tab 98 \tab 73 \tab 86 \tab 1 \tab 41 \tab 213 \tab - \tab - \tab 1606 \cr
   1994 \tab 201 \tab 6 \tab 1457 \tab 53 \tab 253 \tab 36 \tab 196 \tab 130 \tab 185 \tab 4 \tab 112 \tab 358 \tab 1 \tab - \tab 2992 \cr
   1996 \tab 258 \tab 19 \tab 1358 \tab 53 \tab 275 \tab 31 \tab 192 \tab 138 \tab 181 \tab 3 \tab 72 \tab 323 \tab 1 \tab - \tab 2904 \cr
   1998 \tab 383 \tab 34 \tab 1294 \tab 50 \tab 234 \tab 16 \tab 151 \tab 115 \tab 149 \tab 5 \tab 78 \tab 323 \tab - \tab - \tab 2832 \cr
   2000 \tab 417 \tab 19 \tab 1218 \tab 48 \tab 252 \tab 28 \tab 164 \tab 112 \tab 132 \tab 2 \tab 102 \tab 323 \tab - \tab - \tab 2817 \cr
   2002 \tab 489 \tab 5 \tab 1139 \tab 56 \tab 231 \tab 25 \tab 154 \tab 129 \tab 159 \tab 2 \tab 60 \tab 308 \tab 8 \tab - \tab 2765 \cr
   2004 \tab 584 \tab 10 \tab 1225 \tab 47 \tab 248 \tab 28 \tab 126 \tab 113 \tab 117 \tab 6 \tab 46 \tab 253 \tab 9 \tab - \tab 2812 \cr
   2006 \tab 2096 \tab 31 \tab 1256 \tab 59 \tab 247 \tab 23 \tab 162 \tab 109 \tab 126 \tab 7 \tab 43 \tab 348 \tab 3 \tab - \tab 4510 \cr
   2008 \tab 240 \tab 51 \tab 942 \tab 36 \tab 173 \tab 27 \tab 113 \tab 68 \tab 94 \tab 3 \tab 16 \tab 252 \tab 8 \tab - \tab 2023 \cr
   2010 \tab 202 \tab 46 \tab 963 \tab 40 \tab 175 \tab 11 \tab 114 \tab 77 \tab 123 \tab 3 \tab 27 \tab 261 \tab 2 \tab - \tab 2044 \cr
   2012 \tab 238 \tab 60 \tab 898 \tab 34 \tab 170 \tab 13 \tab 94 \tab 69 \tab 121 \tab 4 \tab 28 \tab 245 \tab - \tab - \tab 1974 \cr
   2014 \tab 188 \tab 41 \tab 1272 \tab 45 \tab 225 \tab 23 \tab 145 \tab 93 \tab 138 \tab 2 \tab 52 \tab 311 \tab 3 \tab - \tab 2538 \cr
   2016 \tab 1096 \tab 40 \tab 932 \tab 40 \tab 167 \tab 26 \tab 119 \tab 78 \tab 107 \tab 2 \tab 20 \tab 238 \tab 2 \tab - \tab 2867 \cr
   2018 \tab 942 \tab 33 \tab 737 \tab 35 \tab 145 \tab 6 \tab 90 \tab 58 \tab 64 \tab 1 \tab 17 \tab 219 \tab 1 \tab - \tab 2348 \cr
   2021 \tab 1699 \tab 83 \tab 1416 \tab 40 \tab 151 \tab 25 \tab 83 \tab 64 \tab 109 \tab 7 \tab 12 \tab 336 \tab 7 \tab - \tab 4032 \cr
   2022 \tab 1927 \tab 20 \tab 864 \tab 37 \tab 117 \tab 16 \tab 95 \tab 47 \tab 76 \tab 7 \tab 17 \tab 295 \tab 26 \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 38211 \tab 503 \tab 18112 \tab 718 \tab 3250 \tab 357 \tab 2144 \tab 1501 \tab 2004 \tab 61 \tab 757 \tab 4701 \tab 71 \tab 3309 \tab 75699 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{nummen}},
\code{\link{numwomen}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{sexsex5}}

Other Sex: 
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{nummen}},
\code{\link{numwomen}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{sexsex5}}

Other Split Ballots: 
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}}
}
\concept{High Risk Behaviors}
\concept{Sex}
\concept{Split Ballots}
\keyword{variable}
