% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_553.R
\name{partopen}
\alias{partopen}
\title{Number of sex partner's last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
partopen
}
\details{
Question How many sex partners have you had in the last 12 months?
}
\section{Values}{

\itemize{
\item \code{990} -
\item \code{997} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5062/vshow}{GSS Data Explorer page for \code{partopen}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 14 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab - \tab no answer \tab Total \cr
   1991 \tab 866 \tab 146 \tab 428 \tab 1 \tab 38 \tab 12 \tab 5 \tab 6 \tab 3 \tab 1 \tab 1 \tab 1 \tab 9 \tab 1517 \cr
   Total \tab 866 \tab 146 \tab 428 \tab 1 \tab 38 \tab 12 \tab 5 \tab 6 \tab 3 \tab 1 \tab 1 \tab 1 \tab 9 \tab 1517 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{nummen}},
\code{\link{numwomen}},
\code{\link{partnrs5}},
\code{\link{partopn5}},
\code{\link{sexfreq1}},
\code{\link{sexfreq2}},
\code{\link{sexsex5}}
}
\concept{High Risk Behaviors}
\keyword{variable}
