% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_294.R
\name{partpol}
\alias{partpol}
\title{Political participation important for freedom}
\source{
General Social Survey https://gss.norc.org
}
\description{
partpol
}
\details{
Question I'm going to read you several statements about freedom. Each one may say something true and important about freedom, but I'd like to know how important each statement is to you. Is it one of the most important things about freedom, extremely important, very important, somewhat important, or not too important?
D. Freedom is having the right to participate in politics and elections.
}
\section{Values}{

\itemize{
\item \verb{[1]} one of the most important
\item \verb{[2]} extremely important
\item \verb{[3]} very important
\item \verb{[4]} moderately important
\item \verb{[5]} somewhat important
\item \verb{[6]} not too important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2671/vshow}{GSS Data Explorer page for \code{partpol}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab one of the most important \tab extremely important \tab very important \tab moderately important \tab somewhat important \tab not too important \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 501 \tab 371 \tab 357 \tab 91 \tab 30 \tab 34 \tab 13 \tab 22 \tab 2817 \cr
   Total \tab 1398 \tab 501 \tab 371 \tab 357 \tab 91 \tab 30 \tab 34 \tab 13 \tab 22 \tab 2817 \cr
}
}

\seealso{
Other Freedom: 
\code{\link{choice}},
\code{\link{freenow}},
\code{\link{howfree}},
\code{\link{inpeace}},
\code{\link{leftlone}},
\code{\link{nogovt}},
\code{\link{rfreenow}},
\code{\link{rhowfree}},
\code{\link{satdemoc}}
}
\concept{Freedom}
\keyword{variable}
