% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_327.R
\name{partthon}
\alias{partthon}
\title{Have participated in a walkathon over past 5 years}
\source{
General Social Survey https://gss.norc.org
}
\description{
partthon
}
\details{
Question Over the past 5 years have you done any of the following to express your opinion about an issue or your support for a cause?
G. Participate in a walkathon or marathon to raise money for a cause
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2916/vshow}{GSS Data Explorer page for \code{partthon}}.
}

\seealso{
Other Adult Transitions: 
\code{\link{givchng}},
\code{\link{hlpneedy}}
}
\concept{Adult Transitions}
\keyword{variable}
