% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_463.R
\name{partvol}
\alias{partvol}
\title{In past 12 months, r has participated in charitable or religious volunteer orgs}
\source{
General Social Survey https://gss.norc.org
}
\description{
partvol
}
\details{
Question In the past 12 months, how often, if at all, have you taken part in the activities? Of charitable or religious organisations that do voluntary work?
}
\section{Values}{

\itemize{
\item \code{1} once a week or more
\item \code{2} one to three times a month
\item \code{3} several times a year
\item \code{4} once in the past year
\item \code{5} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7243/vshow}{GSS Data Explorer page for \code{partvol}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab once a week or more \tab one to three times a month \tab several times a year \tab once in the past year \tab never \tab don't know \tab Total \cr
   2018 \tab 1175 \tab 64 \tab 108 \tab 279 \tab 225 \tab 489 \tab 8 \tab 2348 \cr
   Total \tab 1175 \tab 64 \tab 108 \tab 279 \tab 225 \tab 489 \tab 8 \tab 2348 \cr
}
}

\keyword{variable}
