% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_104.R
\name{partyid3}
\alias{partyid3}
\title{Party id of person #3}
\source{
General Social Survey https://gss.norc.org
}
\description{
partyid3
}
\details{
Question Is (NAME) a Democrat, a Republican, an Independent, or what?
3. Party Identification of Third Person
}
\section{Values}{

\itemize{
\item \code{1} democrat
\item \code{2} republican
\item \code{3} independent
\item \code{4} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/957/vshow}{GSS Data Explorer page for \code{partyid3}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab democrat \tab republican \tab independent \tab other \tab don't know \tab no answer \tab Total \cr
   1987 \tab 735 \tab 415 \tab 221 \tab 99 \tab 33 \tab 285 \tab 31 \tab 1819 \cr
   Total \tab 735 \tab 415 \tab 221 \tab 99 \tab 33 \tab 285 \tab 31 \tab 1819 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{partyid1}},
\code{\link{partyid2}},
\code{\link{relig1}},
\code{\link{relig2}},
\code{\link{relig3}},
\code{\link{relig4}},
\code{\link{relig5}},
\code{\link{talkpol1}},
\code{\link{talkpol2}}
}
\concept{Social Networks}
\keyword{variable}
