% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_557.R
\name{parwhere}
\alias{parwhere}
\title{Parents don't know where kids are}
\source{
General Social Survey https://gss.norc.org
}
\description{
parwhere
}
\details{
Question Please tell me whether you strongly agree, agree, disagree, or strongly disagree with the following statements regarding parents:
D. Parents often don''t know where their children are or what they are doing.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5096/vshow}{GSS Data Explorer page for \code{parwhere}}.
}

\seealso{
Other Children: 
\code{\link{parfin}},
\code{\link{pargovt}},
\code{\link{parhardr}},
\code{\link{parrght}},
\code{\link{partaxes}},
\code{\link{partime}},
\code{\link{parwork}},
\code{\link{parworse}},
\code{\link{rolema}}
}
\concept{Children}
\keyword{variable}
