% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_176.R
\name{pastdown}
\alias{pastdown}
\title{Degree national economy has gotten worse}
\source{
General Social Survey https://gss.norc.org
}
\description{
pastdown
}
\details{
Question How about the economy. Would you say that over the past year the nation's economy has gotten better, stayed about the same, or gotten worse?
B. Would you say much worse or somewhat worse?
}
\section{Values}{

\itemize{
\item \verb{[1]} much worse
\item \verb{[2]} somewhat worse
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1643/vshow}{GSS Data Explorer page for \code{pastdown}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab much worse \tab somewhat worse \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2555 \tab 126 \tab 295 \tab 1 \tab 15 \tab 2992 \cr
   Total \tab 2555 \tab 126 \tab 295 \tab 1 \tab 15 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{econfutr}},
\code{\link{futrdown}},
\code{\link{futrup}},
\code{\link{pastup}},
\code{\link{rdiscaff}},
\code{\link{rimmdisc}}
}
\concept{Multiculturalism}
\keyword{variable}
