% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_423.R
\name{pastpay}
\alias{pastpay}
\title{Was r's starting pay higher/lower/same as earning in previous job}
\source{
General Social Survey https://gss.norc.org
}
\description{
pastpay
}
\details{
Question When you went to work for your current employer, was your starting pay higher, lower or about the same as what you were earning in your previous (full time) job?
}
\section{Values}{

\itemize{
\item \code{1} higher
\item \code{2} about the same
\item \code{3} lower
\item \code{4} current employer is first job
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3875/vshow}{GSS Data Explorer page for \code{pastpay}}.
}

\seealso{
Other Global Economics: 
\code{\link{curempyr}},
\code{\link{paychnge}},
\code{\link{whyjbct1}},
\code{\link{whyleave}},
\code{\link{work3yrs}}

Other Income: 
\code{\link{paychnge}}

Other Job: 
\code{\link{curempyr}},
\code{\link{paychnge}},
\code{\link{whyjbct1}},
\code{\link{whyleave}},
\code{\link{work3yrs}}

Other Occupation: 
\code{\link{curempyr}},
\code{\link{paychnge}}
}
\concept{Global Economics}
\concept{Income}
\concept{Job}
\concept{Occupation}
\keyword{variable}
