% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_476.R
\name{pasup}
\alias{pasup}
\title{R's father supervisor}
\source{
General Social Survey https://gss.norc.org
}
\description{
pasup
}
\details{
Question In the job your father had when you were growing up, did your father supervise anyone who was directly responsible to him?
}
\section{Values}{

\itemize{
\item \code{1} no, did not supervise
\item \code{2} yes, supervised 1 to 9 people
\item \code{3} yes, supervised 10 or more
\item \code{4} dk if he supervised anyone
\item \code{5} i dk what my father did-father dead
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4341/vshow}{GSS Data Explorer page for \code{pasup}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab no, did not supervise \tab yes, supervised 1 to 9 people \tab yes, supervised 10 or more \tab dk if he supervised anyone \tab i dk what my father did-father dead \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1545 \tab 447 \tab 321 \tab 189 \tab 172 \tab 70 \tab 2 \tab 71 \tab 2817 \cr
   Total \tab 1545 \tab 447 \tab 321 \tab 189 \tab 172 \tab 70 \tab 2 \tab 71 \tab 2817 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{execrnk}},
\code{\link{famrnk}},
\code{\link{justpay}},
\code{\link{paychild}},
\code{\link{paydowel}},
\code{\link{payhard}},
\code{\link{soctype1}},
\code{\link{soctype2}},
\code{\link{unsklrnk}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
