% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_537.R
\name{patriot2}
\alias{patriot2}
\title{Patriotic feelings lead to intolerance in america}
\source{
General Social Survey https://gss.norc.org
}
\description{
patriot2
}
\details{
Question How much do you agree of disagree that strong patriotic feelings in America…
B. Lead to intolerance in America
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} disagree strongly
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4904/vshow}{GSS Data Explorer page for \code{patriot2}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab no answer \tab Total \cr
   2014 \tab 1264 \tab 86 \tab 443 \tab 311 \tab 275 \tab 49 \tab 106 \tab 4 \tab 2538 \cr
   Total \tab 1264 \tab 86 \tab 443 \tab 311 \tab 275 \tab 49 \tab 106 \tab 4 \tab 2538 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{immassim}},
\code{\link{immcult}},
\code{\link{immeduc}},
\code{\link{letin1a}},
\code{\link{patriot1}},
\code{\link{patriot3}},
\code{\link{patriot4}},
\code{\link{shortcom}}
}
\concept{ISSP National Identity}
\keyword{variable}
