% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_537.R
\name{patriot4}
\alias{patriot4}
\title{Patriotic feelings lead to negative feelings towards immigrants}
\source{
General Social Survey https://gss.norc.org
}
\description{
patriot4
}
\details{
Question How much do you agree of disagree that strong patriotic feelings in America…
D. Lead to negative attitudes towards immigrants in America.
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} disagree strongly
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4906/vshow}{GSS Data Explorer page for \code{patriot4}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{immassim}},
\code{\link{immcult}},
\code{\link{immeduc}},
\code{\link{letin1a}},
\code{\link{patriot1}},
\code{\link{patriot2}},
\code{\link{patriot3}},
\code{\link{shortcom}}

Other Immigration: 
\code{\link{immassim}},
\code{\link{immcult}},
\code{\link{immeduc}},
\code{\link{letin1a}}

Other United States: 
\code{\link{immcult}},
\code{\link{patriot1}},
\code{\link{patriot2}},
\code{\link{patriot3}},
\code{\link{shortcom}}
}
\concept{ISSP National Identity}
\concept{Immigration}
\concept{United States}
\keyword{variable}
