% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_130.R
\name{patrso}
\alias{patrso}
\title{Unpatriotic - patriotic}
\source{
General Social Survey https://gss.norc.org
}
\description{
patrso
}
\details{
Question Now I have some questions about different groups in our society. I'm going to show you a seven-point scale on which the characteristics of people in a group can be rated. In the first statement a score of 1 means that you think almost all of the people in that group are "rich."  A score of 7 means that you think almost everyone in the group are "poor."  A score of 4 means you think that the group is not towards one end or another, and of course you may choose any number In between that comes closest to where you think people in the group stand.
F. Do people in these groups tend to be patriotic or do they tend to be unpatriotic?
6. Southern Whites?
}
\section{Values}{

\itemize{
\item \code{1} patriotic
\item \code{7} unpatriotic
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1213/vshow}{GSS Data Explorer page for \code{patrso}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab patriotic \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab unpatriotic \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 258 \tab 301 \tab 235 \tab 348 \tab 50 \tab 16 \tab 9 \tab 121 \tab 34 \tab 1372 \cr
   Total \tab 0 \tab 258 \tab 301 \tab 235 \tab 348 \tab 50 \tab 16 \tab 9 \tab 121 \tab 34 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{liveasns}},
\code{\link{liveblks}},
\code{\link{livehsps}},
\code{\link{livejews}},
\code{\link{liveno}},
\code{\link{liveso}},
\code{\link{patrasns}},
\code{\link{patrblks}},
\code{\link{patrhsps}}
}
\concept{Intergroup Relations}
\keyword{variable}
