% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_457.R
\name{pavisit1}
\alias{pavisit1}
\title{How often does r visit father}
\source{
General Social Survey https://gss.norc.org
}
\description{
pavisit1
}
\details{
Question And now some questions about your father. How often do you see or visit your father?
}
\section{Values}{

\itemize{
\item \code{1} he lives in the same household as i do
\item \code{2} daily
\item \code{3} at least several times a week
\item \code{4} at least once a week
\item \code{5} at least once a month
\item \code{6} several times a year
\item \code{7} less often
\item \code{8} never
\item \code{9} my father is no longer alive
\item \code{10} i dont know where father lives
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4192/vshow}{GSS Data Explorer page for \code{pavisit1}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab he lives in the same household as i do \tab daily \tab at least several times a week \tab at least once a week \tab at least once a month \tab several times a year \tab less often \tab never \tab my father is no longer alive \tab i dont know where father lives \tab no answer \tab Total \cr
   2002 \tab 1616 \tab 35 \tab 40 \tab 60 \tab 92 \tab 94 \tab 114 \tab 85 \tab 55 \tab 545 \tab 25 \tab 4 \tab 2765 \cr
   Total \tab 1616 \tab 35 \tab 40 \tab 60 \tab 92 \tab 94 \tab 114 \tab 85 \tab 55 \tab 545 \tab 25 \tab 4 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{cousins}},
\code{\link{kidcall}},
\code{\link{macall1}},
\code{\link{mavisit1}},
\code{\link{niecenep}},
\code{\link{pacall1}},
\code{\link{parslaw}},
\code{\link{sibinlaw}},
\code{\link{uncaunts}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
