% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_670.R
\name{pawrkgrw}
\alias{pawrkgrw}
\title{Did father (or oldest same-sex parent) ever work for pay for as long as a year while you were growing up?}
\source{
General Social Survey https://gss.norc.org
}
\description{
pawrkgrw
}
\details{
Question Did father (or oldest same-sex parent) ever work for pay for as long as a year while you were growing up?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8201/vshow}{GSS Data Explorer page for \code{pawrkgrw}}.
}

\seealso{
Other Core: 
\code{\link{conlabory}},
\code{\link{marsamey}},
\code{\link{mobiledata}},
\code{\link{racdif5}}
}
\concept{Core}
\keyword{variable}
