% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_612.R
\name{pawrkslf2}
\alias{pawrkslf2}
\title{At this job, was (dadfill) an employee, self-employed without employees, or self-employed with employees?}
\source{
General Social Survey https://gss.norc.org
}
\description{
pawrkslf2
}
\details{
Question At this job, was your father an employee, self-employed without employees, or self-employed with employees?
}
\section{Values}{

\itemize{
\item \code{1} employee
\item \code{2} self-employed without employees
\item \code{3} self-employed with employees
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7630/vshow}{GSS Data Explorer page for \code{pawrkslf2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab employee \tab self-employed without employees \tab self-employed with employees \tab no answer \tab Total \cr
   2021 \tab 675 \tab 2539 \tab 364 \tab 446 \tab 8 \tab 4032 \cr
   2022 \tab 727 \tab 2197 \tab 303 \tab 314 \tab 3 \tab 3544 \cr
   Total \tab 1402 \tab 4736 \tab 667 \tab 760 \tab 11 \tab 7576 \cr
}
}

\keyword{variable}
