% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_612.R
\name{pawrkslffam}
\alias{pawrkslffam}
\title{Father (or oldest same-sex parent) owned family business or farm}
\source{
General Social Survey https://gss.norc.org
}
\description{
pawrkslffam
}
\details{
Question At this job, was he working for his own family business or family farm?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7429/vshow}{GSS Data Explorer page for \code{pawrkslffam}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab skipped on web \tab Total \cr
   2021 \tab 541 \tab 660 \tab 2813 \tab 5 \tab 13 \tab 4032 \cr
   2022 \tab 727 \tab 549 \tab 2229 \tab 37 \tab 2 \tab 3544 \cr
   Total \tab 1268 \tab 1209 \tab 5042 \tab 42 \tab 15 \tab 7576 \cr
}
}

\keyword{variable}
