% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_470.R
\name{paycabnt}
\alias{paycabnt}
\title{How much r thinks cabinet member makes}
\source{
General Social Survey https://gss.norc.org
}
\description{
paycabnt
}
\details{
Question We would like to know what you think people in these jobs actually earn. Please write in how much you think they usually earn, each year, before taxes. Many people are not exactly sure about this but your best guess will be close enough. This may be difficult, but it is very important. So please try. nPlease write in how much they actually earn  each year.
K. A member of the cabinet in the federal government?
}
\section{Values}{

\itemize{
\item \verb{[1e+06]} $1 million (or more, for 1987 & 2000)
\item \verb{[99999997]} never worked
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   B/C/- \tab 2021 \cr
   Full \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4273/vshow}{GSS Data Explorer page for \code{paycabnt}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 10000 \tab 100500 \tab 101000 \tab 10400 \tab 104000 \tab 106000 \tab 110000 \tab 113000 \tab 11500 \tab 12000 \tab 120000 \tab 125000 \tab 130000 \tab 135000 \tab 14000 \tab 140000 \tab 15000 \tab 150000 \tab 155000 \tab 165000 \tab 17000 \tab 170000 \tab 17500 \tab 175000 \tab 180000 \tab 185000 \tab 19000 \tab 190000 \tab 1e+05 \tab $1 million (or more, for 1987 & 2000) \tab 20000 \tab 22000 \tab 24000 \tab 25000 \tab 250000 \tab 2600 \tab 26000 \tab 28000 \tab 29000 \tab 2e+05 \tab 30000 \tab 30720 \tab 35000 \tab 350000 \tab 38000 \tab 3e+05 \tab 40000 \tab 41600 \tab 416000 \tab 42000 \tab 43000 \tab 45000 \tab 450000 \tab 46000 \tab 48000 \tab 4e+05 \tab 5000 \tab 50000 \tab 52000 \tab 55000 \tab 56000 \tab 57000 \tab 58000 \tab 5e+05 \tab 6000 \tab 60000 \tab 61000 \tab 62000 \tab 62400 \tab 62500 \tab 63000 \tab 64000 \tab 65000 \tab 65500 \tab 68000 \tab 69000 \tab 6e+05 \tab 70000 \tab 71000 \tab 72000 \tab 73000 \tab 75000 \tab 750000 \tab 75200 \tab 76000 \tab 77000 \tab 78000 \tab 79000 \tab 79089 \tab 79800 \tab 8000 \tab 80000 \tab 83000 \tab 84000 \tab 85000 \tab 850000 \tab 86000 \tab 8700 \tab 87500 \tab 88000 \tab 89000 \tab 89500 \tab 90000 \tab 92000 \tab 93000 \tab 9500 \tab 95000 \tab 96000 \tab 98000 \tab can't choose \tab no answer \tab 1000 \tab 105000 \tab 108000 \tab 112000 \tab 115000 \tab 123000 \tab 126000 \tab 138000 \tab 145000 \tab 148000 \tab 160000 \tab 225000 \tab 270000 \tab 3000 \tab 33000 \tab 39000 \tab 400 \tab 41000 \tab 728000 \tab 7e+05 \tab 99000 \tab 0 \tab 1.2e+07 \tab 1.5e+07 \tab 1.8e+07 \tab 100 \tab 100009 \tab 10565600 \tab 107000 \tab 109000 \tab 1234254 \tab 12450000 \tab 1250000 \tab 12500000 \tab 13246 \tab 13463 \tab 1362753 \tab 15 \tab 150 \tab 1500000 \tab 150150 \tab 172000 \tab 174000 \tab 178000 \tab 179000 \tab 18 \tab 182000 \tab 186000 \tab 195000 \tab 196000 \tab 199700 \tab 1e+07 \tab 2.5e+07 \tab 2000 \tab 206000 \tab 207000 \tab 210000 \tab 210500 \tab 215000 \tab 219000 \tab 220000 \tab 223235 \tab 230000 \tab 235000 \tab 235300 \tab 240000 \tab 2499 \tab 2500000 \tab 255000 \tab 260000 \tab 275000 \tab 280 \tab 280000 \tab 285000 \tab 2e+06 \tab 2e+07 \tab 3.5e+07 \tab 30 \tab 318000 \tab 3300000 \tab 33444 \tab 34 \tab 340000 \tab 3600000 \tab 36842 \tab 375000 \tab 3e+06 \tab 3e+07 \tab 4.5e+07 \tab 400099 \tab 430000 \tab 4367372 \tab 4500 \tab 4e+07 \tab 50 \tab 550000 \tab 56778 \tab 59000 \tab 5e+06 \tab 5e+07 \tab 6500 \tab 650000 \tab 6e+06 \tab 7000 \tab 75 \tab 764892 \tab 7891 \tab 87000 \tab 8e+05 \tab 8e+06 \tab 900 \tab 99999 \tab 999999 \tab 9e+05 \tab skipped on web \tab Total \cr
   1987 \tab 255 \tab 5 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 6 \tab 18 \tab 2 \tab 1 \tab 2 \tab 1 \tab 4 \tab 46 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 3 \tab 1 \tab 1 \tab 153 \tab 3 \tab 10 \tab 1 \tab 1 \tab 15 \tab 10 \tab 1 \tab 1 \tab 1 \tab 1 \tab 14 \tab 28 \tab 1 \tab 13 \tab 1 \tab 1 \tab 10 \tab 37 \tab 1 \tab 1 \tab 4 \tab 1 \tab 28 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 104 \tab 2 \tab 11 \tab 2 \tab 1 \tab 1 \tab 8 \tab 2 \tab 104 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 35 \tab 1 \tab 2 \tab 1 \tab 1 \tab 68 \tab 1 \tab 2 \tab 1 \tab 173 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 162 \tab 1 \tab 1 \tab 44 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 7 \tab 1 \tab 55 \tab 3 \tab 1 \tab 1 \tab 11 \tab 1 \tab 3 \tab 69 \tab 202 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   2000 \tab 1545 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 8 \tab - \tab - \tab 2 \tab 29 \tab 56 \tab 9 \tab 3 \tab 1 \tab 11 \tab 4 \tab 107 \tab - \tab 2 \tab - \tab 3 \tab - \tab 12 \tab 8 \tab 1 \tab - \tab 3 \tab 159 \tab 12 \tab 8 \tab - \tab 1 \tab 3 \tab 27 \tab - \tab - \tab - \tab - \tab 76 \tab 9 \tab - \tab 7 \tab 2 \tab - \tab 18 \tab 24 \tab - \tab - \tab - \tab - \tab 10 \tab - \tab - \tab 1 \tab 6 \tab - \tab 50 \tab - \tab 4 \tab - \tab - \tab 1 \tab 14 \tab - \tab 50 \tab - \tab - \tab - \tab - \tab - \tab - \tab 15 \tab - \tab 2 \tab - \tab 1 \tab 32 \tab - \tab - \tab - \tab 66 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab 55 \tab - \tab - \tab 20 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 43 \tab - \tab - \tab - \tab 8 \tab - \tab 1 \tab 90 \tab 164 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 4 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2021 \tab 2029 \tab 9 \tab - \tab - \tab - \tab - \tab - \tab 7 \tab - \tab - \tab - \tab 52 \tab 49 \tab 9 \tab 4 \tab - \tab 13 \tab 2 \tab 194 \tab 1 \tab 5 \tab 1 \tab 12 \tab - \tab 82 \tab 35 \tab 4 \tab 1 \tab 7 \tab 139 \tab 33 \tab 12 \tab - \tab - \tab 2 \tab 178 \tab - \tab - \tab - \tab - \tab 286 \tab 6 \tab - \tab 1 \tab 18 \tab - \tab 92 \tab 4 \tab - \tab - \tab - \tab - \tab 5 \tab 8 \tab - \tab - \tab 47 \tab 1 \tab 22 \tab 1 \tab 3 \tab - \tab - \tab - \tab 78 \tab - \tab 16 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 20 \tab 12 \tab - \tab - \tab - \tab 25 \tab 12 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 41 \tab - \tab - \tab 7 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 21 \tab - \tab - \tab - \tab 5 \tab - \tab - \tab 14 \tab 125 \tab 2 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 3 \tab - \tab 13 \tab 11 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 9 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 6 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 2 \tab 1 \tab 9 \tab 1 \tab 1 \tab 1 \tab 15 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 8 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 17 \tab 1 \tab 1 \tab 1 \tab 1 \tab 9 \tab 52 \tab 4032 \cr
   Total \tab 3829 \tab 17 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 16 \tab 1 \tab 1 \tab 4 \tab 87 \tab 123 \tab 20 \tab 8 \tab 3 \tab 25 \tab 10 \tab 347 \tab 2 \tab 8 \tab 2 \tab 16 \tab 1 \tab 97 \tab 45 \tab 8 \tab 2 \tab 11 \tab 451 \tab 48 \tab 30 \tab 1 \tab 2 \tab 20 \tab 215 \tab 1 \tab 1 \tab 1 \tab 1 \tab 376 \tab 43 \tab 1 \tab 21 \tab 21 \tab 1 \tab 120 \tab 65 \tab 1 \tab 1 \tab 4 \tab 1 \tab 43 \tab 9 \tab 2 \tab 3 \tab 54 \tab 2 \tab 176 \tab 3 \tab 18 \tab 2 \tab 1 \tab 2 \tab 100 \tab 2 \tab 170 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 52 \tab 1 \tab 4 \tab 1 \tab 22 \tab 112 \tab 1 \tab 2 \tab 1 \tab 264 \tab 13 \tab 1 \tab 3 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 258 \tab 1 \tab 1 \tab 71 \tab 2 \tab 2 \tab 1 \tab 1 \tab 3 \tab 9 \tab 1 \tab 119 \tab 3 \tab 1 \tab 1 \tab 24 \tab 1 \tab 4 \tab 173 \tab 491 \tab 4 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 5 \tab 1 \tab 17 \tab 14 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 10 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 6 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 6 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 2 \tab 1 \tab 9 \tab 1 \tab 1 \tab 1 \tab 15 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 8 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 17 \tab 1 \tab 1 \tab 1 \tab 1 \tab 9 \tab 52 \tab 8668 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givmason}},
\code{\link{payfctry}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{payrocc}},
\code{\link{paysales}},
\code{\link{payunskl}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
