% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_476.R
\name{paychild}
\alias{paychild}
\title{How much pay dependent on children}
\source{
General Social Survey https://gss.norc.org
}
\description{
paychild
}
\details{
Question In deciding how much people ought to earn, how important should each of these things be, in your opinion:
E. Whether the person has children to support (how important should that be in deciding pay)?
}
\section{Values}{

\itemize{
\item \code{1} essential
\item \code{2} very important
\item \code{3} fairly important
\item \code{4} not very important
\item \code{5} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   B/C/- \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4332/vshow}{GSS Data Explorer page for \code{paychild}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab essential \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2000 \tab 1545 \tab 163 \tab 352 \tab 266 \tab 249 \tab 134 \tab 56 \tab 52 \tab - \tab 2817 \cr
   2021 \tab 2029 \tab 91 \tab 242 \tab 413 \tab 544 \tab 439 \tab 86 \tab 167 \tab 21 \tab 4032 \cr
   Total \tab 3574 \tab 254 \tab 594 \tab 679 \tab 793 \tab 573 \tab 142 \tab 219 \tab 21 \tab 6849 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{execrnk}},
\code{\link{famrnk}},
\code{\link{justpay}},
\code{\link{pasup}},
\code{\link{paydowel}},
\code{\link{payhard}},
\code{\link{soctype1}},
\code{\link{soctype2}},
\code{\link{unsklrnk}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
