% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_423.R
\name{paychnge}
\alias{paychnge}
\title{How much has r pay changed on r's current job since r began}
\source{
General Social Survey https://gss.norc.org
}
\description{
paychnge
}
\details{
Question Thinking about your current employer, how much has your pay changed on your current job since you began?  would you say…
}
\section{Values}{

\itemize{
\item \code{1} my pay has gone up more than the cost of living
\item \code{2} my pay has stayed about the same as the cost of living
\item \code{3} my pay has not kept up with the cost of living
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3874/vshow}{GSS Data Explorer page for \code{paychnge}}.
}

\seealso{
Other Global Economics: 
\code{\link{agecmeus}},
\code{\link{autonojb}},
\code{\link{bornsp}},
\code{\link{curempyr}},
\code{\link{emphlth}},
\code{\link{emphplan}},
\code{\link{findnwjb}},
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{moretrde}},
\code{\link{natdeg}},
\code{\link{newjobs}},
\code{\link{othplan}},
\code{\link{pastpay}},
\code{\link{typevisa}},
\code{\link{visa}},
\code{\link{whyjbct1}},
\code{\link{whyjbct2}},
\code{\link{whyleave}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wkcomptr}},
\code{\link{wocomptr}},
\code{\link{work3yrs}}
}
\concept{Global Economics}
\keyword{variable}
