% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_476.R
\name{paydowel}
\alias{paydowel}
\title{How much pay dependent on work well done}
\source{
General Social Survey https://gss.norc.org
}
\description{
paydowel
}
\details{
Question In deciding how much people ought to earn, how important should each of these things be, in your opinion:
F. How well he or she does the job (how important should that be)?
}
\section{Values}{

\itemize{
\item \code{1} essential
\item \code{2} very important
\item \code{3} fairly important
\item \code{4} not very important
\item \code{5} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   B/C/- \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4333/vshow}{GSS Data Explorer page for \code{paydowel}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab essential \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2000 \tab 1545 \tab 484 \tab 618 \tab 83 \tab 10 \tab 6 \tab 21 \tab 50 \tab - \tab 2817 \cr
   2021 \tab 2029 \tab 763 \tab 796 \tab 171 \tab 24 \tab 6 \tab 55 \tab 166 \tab 22 \tab 4032 \cr
   Total \tab 3574 \tab 1247 \tab 1414 \tab 254 \tab 34 \tab 12 \tab 76 \tab 216 \tab 22 \tab 6849 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{execrnk}},
\code{\link{famrnk}},
\code{\link{justpay}},
\code{\link{pasup}},
\code{\link{paychild}},
\code{\link{payhard}},
\code{\link{soctype1}},
\code{\link{soctype2}},
\code{\link{unsklrnk}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
