% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_476.R
\name{paydowel}
\alias{paydowel}
\title{How much pay dependent on work well done}
\source{
General Social Survey https://gss.norc.org
}
\description{
paydowel
}
\details{
Question In deciding how much people ought to earn, how important should each of these things be, in your opinion:
F. How well he or she does the job (how important should that be)?
}
\section{Values}{

\itemize{
\item \code{1} essential
\item \code{2} very important
\item \code{3} fairly important
\item \code{4} not very important
\item \code{5} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2021 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4333/vshow}{GSS Data Explorer page for \code{paydowel}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{execrnk}},
\code{\link{famrnk}},
\code{\link{justpay}},
\code{\link{pasup}},
\code{\link{paychild}},
\code{\link{payhard}},
\code{\link{soctype1}},
\code{\link{soctype2}},
\code{\link{unsklrnk}}

Other Income: 
\code{\link{justpay}},
\code{\link{paychild}},
\code{\link{payhard}}
}
\concept{ISSP Social Inequality}
\concept{Income}
\keyword{variable}
