% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_503.R
\name{payfam}
\alias{payfam}
\title{How important family responsibility in deciding p}
\source{
General Social Survey https://gss.norc.org
}
\description{
payfam
}
\details{
Question In deciding on pay for two people doing the same kind of work, how important should be …
B. The person''s family responsibilities?
}
\section{Values}{

\itemize{
\item \verb{[1]} essential
\item \verb{[2]} very important
\item \verb{[3]} fairly important
\item \verb{[4]} not very important
\item \verb{[5]} not important at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4572/vshow}{GSS Data Explorer page for \code{payfam}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab essential \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1609 \tab 82 \tab 252 \tab 287 \tab 324 \tab 209 \tab 29 \tab 40 \tab 2832 \cr
   Total \tab 1609 \tab 82 \tab 252 \tab 287 \tab 324 \tab 209 \tab 29 \tab 40 \tab 2832 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{paytime}},
\code{\link{placewrk}},
\code{\link{sethours}},
\code{\link{techjobs}},
\code{\link{techwork}},
\code{\link{wantjob1}},
\code{\link{worknow}}
}
\concept{Work Orientation}
\keyword{variable}
