% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_475.R
\name{payfam1}
\alias{payfam1}
\title{How much pay dependent on family to support}
\source{
General Social Survey https://gss.norc.org
}
\description{
payfam1
}
\details{
Question In deciding how much people ought to earn, how important should each of these things be, in your opinion:
D. What is needed to support a family?
}
\section{Values}{

\itemize{
\item \code{1} essential
\item \code{2} very important
\item \code{3} fairly important
\item \code{4} not very important
\item \code{5} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4331/vshow}{GSS Data Explorer page for \code{payfam1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab essential \tab very important \tab fairly important \tab not very important \tab not important at all \tab can't choose \tab no answer \tab Total \cr
   2000 \tab 1545 \tab 263 \tab 456 \tab 296 \tab 109 \tab 55 \tab 42 \tab 51 \tab 2817 \cr
   Total \tab 1545 \tab 263 \tab 456 \tab 296 \tab 109 \tab 55 \tab 42 \tab 51 \tab 2817 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{corrupt}},
\code{\link{earndes}},
\code{\link{ldcgap}},
\code{\link{ldctax}},
\code{\link{payedtrn}},
\code{\link{payresp}},
\code{\link{paysup}},
\code{\link{richeduc}},
\code{\link{richhlth}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
