% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_154.R
\name{payrborn}
\alias{payrborn}
\title{Yr r father born}
\source{
General Social Survey https://gss.norc.org
}
\description{
payrborn
}
\details{
Question In what year was your father born?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1458/vshow}{GSS Data Explorer page for \code{payrborn}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1857 \tab 1860 \tab 1864 \tab 1865 \tab 1866 \tab 1869 \tab 1870 \tab 1871 \tab 1872 \tab 1873 \tab 1874 \tab 1875 \tab 1876 \tab 1877 \tab 1878 \tab 1879 \tab 1880 \tab 1881 \tab 1882 \tab 1883 \tab 1884 \tab 1885 \tab 1886 \tab 1887 \tab 1888 \tab 1889 \tab 1890 \tab 1891 \tab 1892 \tab 1893 \tab 1894 \tab 1895 \tab 1896 \tab 1897 \tab 1898 \tab 1899 \tab 1900 \tab 1901 \tab 1902 \tab 1903 \tab 1904 \tab 1905 \tab 1906 \tab 1907 \tab 1908 \tab 1909 \tab 1910 \tab 1911 \tab 1912 \tab 1913 \tab 1914 \tab 1915 \tab 1916 \tab 1917 \tab 1918 \tab 1919 \tab 1920 \tab 1921 \tab 1922 \tab 1923 \tab 1924 \tab 1925 \tab 1926 \tab 1927 \tab 1928 \tab 1929 \tab 1930 \tab 1931 \tab 1932 \tab 1933 \tab 1934 \tab 1935 \tab 1936 \tab 1937 \tab 1938 \tab 1939 \tab 1940 \tab 1941 \tab 1942 \tab 1943 \tab 1944 \tab 1945 \tab 1946 \tab 1947 \tab 1948 \tab 1949 \tab 1950 \tab 1951 \tab 1952 \tab 1953 \tab 1954 \tab 1955 \tab 1957 \tab 1958 \tab 1960 \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1474 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 5 \tab 1 \tab 2 \tab 3 \tab 4 \tab 4 \tab 1 \tab 2 \tab 6 \tab 4 \tab 5 \tab 3 \tab 9 \tab 11 \tab 13 \tab 8 \tab 12 \tab 4 \tab 13 \tab 8 \tab 9 \tab 6 \tab 10 \tab 13 \tab 9 \tab 11 \tab 14 \tab 10 \tab 14 \tab 11 \tab 15 \tab 9 \tab 13 \tab 18 \tab 6 \tab 19 \tab 16 \tab 17 \tab 27 \tab 13 \tab 17 \tab 25 \tab 32 \tab 14 \tab 28 \tab 26 \tab 20 \tab 36 \tab 27 \tab 26 \tab 22 \tab 29 \tab 26 \tab 32 \tab 28 \tab 27 \tab 23 \tab 31 \tab 27 \tab 27 \tab 25 \tab 23 \tab 27 \tab 22 \tab 27 \tab 22 \tab 23 \tab 18 \tab 17 \tab 21 \tab 19 \tab 11 \tab 20 \tab 22 \tab 6 \tab 12 \tab 8 \tab 9 \tab 7 \tab 2 \tab 3 \tab 3 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 237 \tab 12 \tab 2992 \cr
   Total \tab 1474 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 5 \tab 1 \tab 2 \tab 3 \tab 4 \tab 4 \tab 1 \tab 2 \tab 6 \tab 4 \tab 5 \tab 3 \tab 9 \tab 11 \tab 13 \tab 8 \tab 12 \tab 4 \tab 13 \tab 8 \tab 9 \tab 6 \tab 10 \tab 13 \tab 9 \tab 11 \tab 14 \tab 10 \tab 14 \tab 11 \tab 15 \tab 9 \tab 13 \tab 18 \tab 6 \tab 19 \tab 16 \tab 17 \tab 27 \tab 13 \tab 17 \tab 25 \tab 32 \tab 14 \tab 28 \tab 26 \tab 20 \tab 36 \tab 27 \tab 26 \tab 22 \tab 29 \tab 26 \tab 32 \tab 28 \tab 27 \tab 23 \tab 31 \tab 27 \tab 27 \tab 25 \tab 23 \tab 27 \tab 22 \tab 27 \tab 22 \tab 23 \tab 18 \tab 17 \tab 21 \tab 19 \tab 11 \tab 20 \tab 22 \tab 6 \tab 12 \tab 8 \tab 9 \tab 7 \tab 2 \tab 3 \tab 3 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 237 \tab 12 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{malive1}},
\code{\link{mayrborn}},
\code{\link{mayrdied}},
\code{\link{palive1}}
}
\concept{Family Mobility}
\keyword{variable}
