% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_154.R
\name{payrborn}
\alias{payrborn}
\title{Yr r father born}
\source{
General Social Survey https://gss.norc.org
}
\description{
payrborn
}
\details{
Question In what year was your father born?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1458/vshow}{GSS Data Explorer page for \code{payrborn}}.
}

\seealso{
Other Family Mobility: 
\code{\link{malive1}},
\code{\link{mayrborn}},
\code{\link{mayrdied}},
\code{\link{palive1}}

Other Date Of Birth: 
\code{\link{mayrborn}}

Other Family: 
\code{\link{malive1}},
\code{\link{mayrborn}},
\code{\link{mayrdied}},
\code{\link{palive1}}

Other Parents: 
\code{\link{malive1}},
\code{\link{mayrborn}},
\code{\link{mayrdied}},
\code{\link{palive1}}
}
\concept{Date Of Birth}
\concept{Family}
\concept{Family Mobility}
\concept{Parents}
\keyword{variable}
