% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_475.R
\name{paysup}
\alias{paysup}
\title{How much pay dependent on supervising others}
\source{
General Social Survey https://gss.norc.org
}
\description{
paysup
}
\details{
Question In deciding how much people ought to earn, how important should each of these things be, in your opinion:
C. Whether the job requires supervising others (how important should that be in deciding pay)?
}
\section{Values}{

\itemize{
\item \code{1} essential
\item \code{2} very important
\item \code{3} fairly important
\item \code{4} not very important
\item \code{5} not important at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4330/vshow}{GSS Data Explorer page for \code{paysup}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{corrupt}},
\code{\link{earndes}},
\code{\link{ldcgap}},
\code{\link{ldctax}},
\code{\link{payedtrn}},
\code{\link{payfam1}},
\code{\link{payresp}},
\code{\link{richeduc}},
\code{\link{richhlth}}

Other Income: 
\code{\link{earndes}},
\code{\link{payedtrn}},
\code{\link{payfam1}},
\code{\link{payresp}}
}
\concept{ISSP Social Inequality}
\concept{Income}
\keyword{variable}
