% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_538.R
\name{paytaxes}
\alias{paytaxes}
\title{How important never to try to evade taxes}
\source{
General Social Survey https://gss.norc.org
}
\description{
paytaxes
}
\details{
Question There are different opinions as to what it takes to be a good citizen. As far as you are concerned personally on a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it…
B. Never to try to evade taxes
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all important
\item \verb{[7]} very important
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4908/vshow}{GSS Data Explorer page for \code{paytaxes}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab not at all important \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 28 \tab 9 \tab 22 \tab 64 \tab 107 \tab 165 \tab 1066 \tab 7 \tab 4 \tab 2812 \cr
   2014 \tab 1274 \tab 25 \tab 6 \tab 13 \tab 50 \tab 67 \tab 114 \tab 955 \tab 31 \tab 3 \tab 2538 \cr
   Total \tab 2614 \tab 53 \tab 15 \tab 35 \tab 114 \tab 174 \tab 279 \tab 2021 \tab 38 \tab 7 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actassoc}},
\code{\link{buypol}},
\code{\link{helpusa}},
\code{\link{helpwrld}},
\code{\link{milserve}},
\code{\link{obeylaws}},
\code{\link{othreasn}},
\code{\link{relmeet}},
\code{\link{watchgov}}
}
\concept{ISSP Citizenship}
\keyword{variable}
