% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_470.R
\name{payunskl}
\alias{payunskl}
\title{How much r thinks unskilled wrkr makes}
\source{
General Social Survey https://gss.norc.org
}
\description{
payunskl
}
\details{
Question We would like to know what you think people in these jobs actually earn. Please write in how much you think they usually earn, each year, before taxes. Many people are not exactly sure about this but your best guess will be close enough. This may be difficult, but it is very important. So please try. nPlease write in how much they actually earn  each year.
J. An unskilled worker in a factory?
}
\section{Values}{

\itemize{
\item \code{1e+06} $1 million (or more, for 1987 & 2000)
\item \code{99999997} never worked
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   B/C/- \tab 2021 \cr
   Full \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4272/vshow}{GSS Data Explorer page for \code{payunskl}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1000 \tab 10000 \tab 10296 \tab 10400 \tab 1050 \tab 10500 \tab 11000 \tab 11500 \tab 1200 \tab 12000 \tab 12480 \tab 12500 \tab 13000 \tab 13500 \tab 14000 \tab 14400 \tab 14560 \tab 1500 \tab 15000 \tab 15500 \tab 15600 \tab 16000 \tab 16640 \tab 1700 \tab 17000 \tab 17500 \tab 1800 \tab 18000 \tab 18500 \tab 19000 \tab $1 million (or more, for 1987 & 2000) \tab 20000 \tab 20800 \tab 21000 \tab 22000 \tab 23000 \tab 24000 \tab 24960 \tab 2500 \tab 25000 \tab 26000 \tab 27000 \tab 27500 \tab 2800 \tab 28000 \tab 29000 \tab 3000 \tab 30000 \tab 31000 \tab 3120 \tab 31200 \tab 3200 \tab 32000 \tab 3500 \tab 35000 \tab 38000 \tab 4000 \tab 40000 \tab 47000 \tab 4800 \tab 5000 \tab 5800 \tab 6000 \tab 6432 \tab 6500 \tab 6968 \tab 7000 \tab 70000 \tab 7250 \tab 7280 \tab 7500 \tab 75000 \tab 7600 \tab 7800 \tab 8000 \tab 8200 \tab 8500 \tab 900 \tab 9000 \tab 9500 \tab 9600 \tab 9950 \tab can't choose \tab no answer \tab 140000 \tab 150000 \tab 18050 \tab 2200 \tab 2e+05 \tab 300 \tab 37000 \tab 416 \tab 44000 \tab 45000 \tab 50000 \tab 52000 \tab 6 \tab 60000 \tab 6e+05 \tab 80000 \tab 0 \tab 11 \tab 12 \tab 120000 \tab 1223 \tab 123445 \tab 125000 \tab 12960 \tab 14 \tab 14500 \tab 15 \tab 1500000 \tab 16 \tab 1600 \tab 17160 \tab 180000 \tab 190000 \tab 1900000 \tab 1e+05 \tab 20 \tab 200 \tab 2000 \tab 20900 \tab 21 \tab 22500 \tab 2300 \tab 23500 \tab 23999 \tab 250000 \tab 25002 \tab 25500 \tab 25680 \tab 2700 \tab 2900 \tab 29120 \tab 2e+06 \tab 30896 \tab 32640 \tab 33000 \tab 34000 \tab 35 \tab 350000 \tab 3500000 \tab 36000 \tab 37500 \tab 39000 \tab 3e+05 \tab 42000 \tab 422222 \tab 43000 \tab 450000 \tab 46000 \tab 48000 \tab 49000 \tab 50 \tab 500 \tab 53000 \tab 55000 \tab 5e+05 \tab 600 \tab 65000 \tab 650000 \tab 7 \tab 75 \tab 750 \tab 7555 \tab 800 \tab 8902 \tab skipped on web \tab Total \cr
   1987 \tab 255 \tab 6 \tab 200 \tab 1 \tab 2 \tab 1 \tab 1 \tab 29 \tab 2 \tab 1 \tab 140 \tab 1 \tab 6 \tab 50 \tab 2 \tab 40 \tab 1 \tab 1 \tab 2 \tab 178 \tab 1 \tab 1 \tab 39 \tab 2 \tab 1 \tab 25 \tab 3 \tab 1 \tab 77 \tab 2 \tab 12 \tab 1 \tab 142 \tab 1 \tab 6 \tab 14 \tab 2 \tab 5 \tab 1 \tab 1 \tab 44 \tab 1 \tab 2 \tab 2 \tab 1 \tab 5 \tab 1 \tab 5 \tab 20 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 8 \tab 5 \tab 1 \tab 1 \tab 22 \tab 1 \tab 23 \tab 1 \tab 4 \tab 6 \tab 43 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 57 \tab 1 \tab 6 \tab 1 \tab 41 \tab 2 \tab 1 \tab 1 \tab 59 \tab 179 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   2000 \tab 1545 \tab 3 \tab 83 \tab - \tab 2 \tab - \tab - \tab 5 \tab - \tab - \tab 66 \tab - \tab - \tab 15 \tab - \tab 15 \tab - \tab 1 \tab - \tab 163 \tab 1 \tab - \tab 34 \tab - \tab - \tab 16 \tab - \tab - \tab 61 \tab - \tab 11 \tab 1 \tab 204 \tab - \tab 2 \tab 18 \tab 6 \tab 6 \tab - \tab - \tab 104 \tab 1 \tab - \tab - \tab - \tab 8 \tab 3 \tab - \tab 79 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 24 \tab - \tab 2 \tab 26 \tab - \tab - \tab 5 \tab - \tab 7 \tab - \tab - \tab - \tab 11 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 13 \tab - \tab - \tab - \tab 10 \tab 1 \tab - \tab - \tab 63 \tab 172 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 8 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2021 \tab 2029 \tab 8 \tab 31 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 11 \tab - \tab - \tab 4 \tab - \tab 3 \tab - \tab - \tab 1 \tab 65 \tab - \tab - \tab 19 \tab - \tab - \tab 8 \tab - \tab - \tab 26 \tab - \tab 12 \tab - \tab 183 \tab - \tab 5 \tab 24 \tab 13 \tab 21 \tab 1 \tab - \tab 244 \tab 8 \tab 9 \tab 1 \tab - \tab 40 \tab 1 \tab 3 \tab 319 \tab 4 \tab - \tab - \tab - \tab 25 \tab - \tab 162 \tab 10 \tab - \tab 189 \tab 2 \tab - \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 7 \tab 5 \tab - \tab - \tab - \tab 11 \tab - \tab - \tab 3 \tab - \tab - \tab 1 \tab 6 \tab - \tab - \tab - \tab 11 \tab 125 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 3 \tab - \tab 3 \tab 65 \tab 77 \tab - \tab - \tab 31 \tab 1 \tab 4 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 5 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 5 \tab 1 \tab 2 \tab 2 \tab 12 \tab 2 \tab 2 \tab 3 \tab 4 \tab 1 \tab 2 \tab 2 \tab 4 \tab 4 \tab 1 \tab 2 \tab 2 \tab 2 \tab 14 \tab 2 \tab 2 \tab 8 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 43 \tab 4032 \cr
   Total \tab 3829 \tab 17 \tab 314 \tab 1 \tab 4 \tab 1 \tab 1 \tab 34 \tab 2 \tab 3 \tab 217 \tab 1 \tab 6 \tab 69 \tab 2 \tab 58 \tab 1 \tab 2 \tab 3 \tab 406 \tab 2 \tab 1 \tab 92 \tab 2 \tab 1 \tab 49 \tab 3 \tab 1 \tab 164 \tab 2 \tab 35 \tab 2 \tab 529 \tab 1 \tab 13 \tab 56 \tab 21 \tab 32 \tab 2 \tab 1 \tab 392 \tab 10 \tab 11 \tab 3 \tab 1 \tab 53 \tab 5 \tab 8 \tab 418 \tab 5 \tab 1 \tab 1 \tab 1 \tab 28 \tab 3 \tab 188 \tab 11 \tab 10 \tab 220 \tab 3 \tab 1 \tab 30 \tab 1 \tab 31 \tab 1 \tab 4 \tab 6 \tab 61 \tab 7 \tab 1 \tab 2 \tab 2 \tab 14 \tab 1 \tab 1 \tab 73 \tab 1 \tab 6 \tab 2 \tab 57 \tab 3 \tab 1 \tab 1 \tab 133 \tab 476 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 4 \tab 1 \tab 4 \tab 66 \tab 85 \tab 1 \tab 1 \tab 33 \tab 2 \tab 5 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 5 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 5 \tab 1 \tab 2 \tab 2 \tab 12 \tab 2 \tab 2 \tab 3 \tab 4 \tab 1 \tab 2 \tab 2 \tab 4 \tab 4 \tab 1 \tab 2 \tab 2 \tab 2 \tab 14 \tab 2 \tab 2 \tab 8 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 43 \tab 8668 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{givclerk}},
\code{\link{givdoc}},
\code{\link{givmason}},
\code{\link{paycabnt}},
\code{\link{payfctry}},
\code{\link{payjudge}},
\code{\link{paylaw}},
\code{\link{payrocc}},
\code{\link{paysales}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
