% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_280.R
\name{paywww}
\alias{paywww}
\title{Does employer pay for www}
\source{
General Social Survey https://gss.norc.org
}
\description{
paywww
}
\details{
Question B. Does your employer provide all or part of your Internet access from home?
}
\section{Values}{

\itemize{
\item \code{1} employer provides internet connection at no charge
\item \code{2} covers some but not all of cost
\item \code{3} does not contribute to cost of using internet
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2553/vshow}{GSS Data Explorer page for \code{paywww}}.
}

\seealso{
Other Information Society: 
\code{\link{comphome}},
\code{\link{ecomhrh}},
\code{\link{ecomminh}},
\code{\link{numhome}},
\code{\link{perminh}},
\code{\link{usehome}},
\code{\link{wwwhome}},
\code{\link{wwwline1}},
\code{\link{wwwline2}}
}
\concept{Information Society}
\keyword{variable}
