% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_151.R
\name{pclit}
\alias{pclit}
\title{Shame tradl amer lit ignored for minority lit}
\source{
General Social Survey https://gss.norc.org
}
\description{
pclit
}
\details{
Question I'm going to read several statements. After I read each one, please tell me whether you strongly agree, agree, disagree, or strongly disagree with it. For example, ...
G. It is a shame when traditional American literature is ignored while other works are promoted because they are by women or by members of minority groups.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1428/vshow}{GSS Data Explorer page for \code{pclit}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 115 \tab 834 \tab 412 \tab 58 \tab 172 \tab 15 \tab 1606 \cr
   1998 \tab 1387 \tab 129 \tab 546 \tab 446 \tab 88 \tab 213 \tab 23 \tab 2832 \cr
   Total \tab 1387 \tab 244 \tab 1380 \tab 858 \tab 146 \tab 385 \tab 38 \tab 4438 \cr
}
}

\seealso{
Other Culture: 
\code{\link{classics}},
\code{\link{english}},
\code{\link{grtbooks}},
\code{\link{judgeart}},
\code{\link{modpaint}},
\code{\link{trstprof}},
\code{\link{tvnews}},
\code{\link{tvpbs}},
\code{\link{tvshows}}
}
\concept{Culture}
\keyword{variable}
