% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_296.R
\name{peaceful}
\alias{peaceful}
\title{Felt peaceful in past 4 wks}
\source{
General Social Survey https://gss.norc.org
}
\description{
peaceful
}
\details{
Question During the past 4 weeks, have you felt calm and peaceful?
}
\section{Values}{

\itemize{
\item \code{1} all of the time
\item \code{2} most of the time
\item \code{3} a good bit of the time
\item \code{4} some of the time
\item \code{5} a little bit of the time
\item \code{6} non of the time
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2684/vshow}{GSS Data Explorer page for \code{peaceful}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab all of the time \tab most of the time \tab a good bit of the time \tab some of the time \tab a little bit of the time \tab non of the time \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 144 \tab 607 \tab 252 \tab 280 \tab 93 \tab 32 \tab 11 \tab 2817 \cr
   Total \tab 1398 \tab 144 \tab 607 \tab 252 \tab 280 \tab 93 \tab 32 \tab 11 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{downblue}},
\code{\link{energy}},
\code{\link{pain}},
\code{\link{socacts}},
\code{\link{treat1}},
\code{\link{treat2}},
\code{\link{treat3}},
\code{\link{treat4}},
\code{\link{treat5}}
}
\concept{Health Status}
\keyword{variable}
