% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_450.R
\name{peocntct}
\alias{peocntct}
\title{How many people in contact in a typical weekday}
\source{
General Social Survey https://gss.norc.org
}
\description{
peocntct
}
\details{
Question On average, about how many people do you have contact with in a typical week day, including people you live with.
We are interested in contact on a  one-to-one basis, including everyone with whom you chat, talk, or discuss matters.
This can be face-to-face, by telephone, by mail, or on the internet. Please include only people you know.
Please tell me which of the following categories best matches your estimate.
Would you say…
}
\section{Values}{

\itemize{
\item \code{1} 0-4 persons
\item \code{2} 5-9
\item \code{3} 10-19
\item \code{4} 20-49
\item \code{5} 50 or more
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D/-/- \tab partial \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4126/vshow}{GSS Data Explorer page for \code{peocntct}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{macall}},
\code{\link{malive}},
\code{\link{matime}},
\code{\link{mavisit}},
\code{\link{palive}}
}
\concept{ISSP Social Networks & Support Systems}
\concept{Social Networks}
\keyword{variable}
