% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_528.R
\name{peopgrn}
\alias{peopgrn}
\title{How get people to protect environment}
\source{
General Social Survey https://gss.norc.org
}
\description{
peopgrn
}
\details{
Question J. Which of these approaches do you think would be the best way of getting people and their families in America to protect the environment?
}
\section{Values}{

\itemize{
\item \code{1} heavy fines for people who damage the environment
\item \code{2} use the tax system to reward people who protect the environment
\item \code{3} more information and education for people about the advantages of protecting the environment
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2010, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4812/vshow}{GSS Data Explorer page for \code{peopgrn}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab heavy fines for people who damage the environment \tab use the tax system to reward people who protect the environment \tab more information and education for people about the advantages of protecting the environment \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2010 \tab 614 \tab 206 \tab 441 \tab 676 \tab 102 \tab 5 \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 228 \tab 727 \tab 734 \tab 132 \tab 191 \tab 17 \tab 4032 \cr
   Total \tab 2617 \tab 434 \tab 1168 \tab 1410 \tab 234 \tab 196 \tab 17 \tab 6076 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{busgrn}},
\code{\link{grneffme}},
\code{\link{h2oless}},
\code{\link{helpharm}},
\code{\link{nobuygrn}},
\code{\link{redcehme}},
\code{\link{tempgen1}}
}
\concept{ISSP Environment}
\keyword{variable}
