% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_652.R
\name{perfrt}
\alias{perfrt}
\title{Performance rating last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
perfrt
}
\details{
Question If you received a performance rating from your organization in 2021, which statement best describes the rating you received?
}
\section{Values}{

\itemize{
\item \verb{[1]} below expectations
\item \verb{[2]} meets expectations
\item \verb{[3]} above expectations
\item \verb{[4]} r didn't receive a performance rating
\item \verb{[5]} none of the statements match the performance rating r received
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7853/vshow}{GSS Data Explorer page for \code{perfrt}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab below expectations \tab meets expectations \tab above expectations \tab r didn't receive a performance rating \tab none of the statements match the performance rating r received \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2569 \tab 26 \tab 216 \tab 364 \tab 264 \tab 37 \tab 48 \tab 8 \tab 12 \tab 3544 \cr
   Total \tab 2569 \tab 26 \tab 216 \tab 364 \tab 264 \tab 37 \tab 48 \tab 8 \tab 12 \tab 3544 \cr
}
}

\keyword{variable}
