% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_518.R
\name{perscrfc}
\alias{perscrfc}
\title{R made some personal sacrifice as an expression of faith by fasting, giving up a}
\source{
General Social Survey https://gss.norc.org
}
\description{
perscrfc
}
\details{
Question During the last year, did you make some personal sacrifice as an expression of your faith such as by fasting, following a special diet, or giving up some activity during a holy season such as Lent or Ramadan?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4716/vshow}{GSS Data Explorer page for \code{perscrfc}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2008 \tab 658 \tab 360 \tab 993 \tab 5 \tab 7 \tab 2023 \cr
   Total \tab 658 \tab 360 \tab 993 \tab 5 \tab 7 \tab 2023 \cr
}
}

\concept{ISSP Religion}
\keyword{variable}
