% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_339.R
\name{pessimst}
\alias{pessimst}
\title{I hardly ever expect things to go my way}
\source{
General Social Survey https://gss.norc.org
}
\description{
pessimst
}
\details{
Question Indicate your agreement with each of the following statements by selecting the number that comes closest to your answer:
G. I hardly ever expect things to go my way.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3042/vshow}{GSS Data Explorer page for \code{pessimst}}.
}

\seealso{
Other Work Organizations: 
\code{\link{afailure}},
\code{\link{nogood}},
\code{\link{notcount}},
\code{\link{ofworth}},
\code{\link{optimist}},
\code{\link{satself}},
\code{\link{slfrspct}}

Other Genes and the Environment: 
\code{\link{afailure}},
\code{\link{genenvo3}},
\code{\link{genenvo4}},
\code{\link{nogood}},
\code{\link{notcount}},
\code{\link{ofworth}},
\code{\link{optimist}},
\code{\link{satself}},
\code{\link{slfrspct}}
}
\concept{Genes and the Environment}
\concept{Work Organizations}
\keyword{variable}
