% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_317.R
\name{petb4}
\alias{petb4}
\title{Did r's family have a pet 5 years ago}
\source{
General Social Survey https://gss.norc.org
}
\description{
petb4
}
\details{
Question Did your family have a pet 5 years ago?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7196/vshow}{GSS Data Explorer page for \code{petb4}}.
}

\keyword{variable}
