% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_320.R
\name{petb4cmfrt}
\alias{petb4cmfrt}
\title{How often r used to seek pet for comfort when feeling bad}
\source{
General Social Survey https://gss.norc.org
}
\description{
petb4cmfrt
}
\details{
Question When you felt bad, how often did you seek your pet for comfort?
}
\section{Values}{

\itemize{
\item \verb{[1]} almost always
\item \verb{[2]} often
\item \verb{[3]} sometimes
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7221/vshow}{GSS Data Explorer page for \code{petb4cmfrt}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab almost always \tab often \tab sometimes \tab never \tab no answer \tab Total \cr
   2018 \tab 2144 \tab 55 \tab 43 \tab 49 \tab 34 \tab 23 \tab 2348 \cr
   Total \tab 2144 \tab 55 \tab 43 \tab 49 \tab 34 \tab 23 \tab 2348 \cr
}
}

\keyword{variable}
