% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_320.R
\name{petb4cmfrt}
\alias{petb4cmfrt}
\title{How often r used to seek pet for comfort when feeling bad}
\source{
General Social Survey https://gss.norc.org
}
\description{
petb4cmfrt
}
\details{
Question When you felt bad, how often did you seek your pet for comfort?
}
\section{Values}{

\itemize{
\item \code{1} almost always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7221/vshow}{GSS Data Explorer page for \code{petb4cmfrt}}.
}

\keyword{variable}
