% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_319.R
\name{petcmfrt}
\alias{petcmfrt}
\title{How often r seeks pet for comfort when feeling bad}
\source{
General Social Survey https://gss.norc.org
}
\description{
petcmfrt
}
\details{
Question When you feel bad, how often do you seek your pet for comfort?
}
\section{Values}{

\itemize{
\item \code{1} almost always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7208/vshow}{GSS Data Explorer page for \code{petcmfrt}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab almost always \tab often \tab sometimes \tab never \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1649 \tab 228 \tab 166 \tab 156 \tab 120 \tab 2 \tab 27 \tab 2348 \cr
   Total \tab 1649 \tab 228 \tab 166 \tab 156 \tab 120 \tab 2 \tab 27 \tab 2348 \cr
}
}

\keyword{variable}
