% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_319.R
\name{petfam}
\alias{petfam}
\title{How often r considers pet member of family}
\source{
General Social Survey https://gss.norc.org
}
\description{
petfam
}
\details{
Question How often do you consider your pet to be a member of your family?
}
\section{Values}{

\itemize{
\item \code{1} almost always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7209/vshow}{GSS Data Explorer page for \code{petfam}}.
}

\keyword{variable}
