% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_269.R
\name{petition}
\alias{petition}
\title{Web make r sign petition}
\source{
General Social Survey https://gss.norc.org
}
\description{
petition
}
\details{
Question Have you ever taken any of the following actions because of something you learned about or discussed on the Web?
A. Signed a petition?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2442/vshow}{GSS Data Explorer page for \code{petition}}.
}

\seealso{
Other Information Society: 
\code{\link{contact}},
\code{\link{meeting}},
\code{\link{polchnge}},
\code{\link{polconf}},
\code{\link{poldsagr}},
\code{\link{polforms}},
\code{\link{polneutl}},
\code{\link{polnew}},
\code{\link{votefor}}
}
\concept{Information Society}
\keyword{variable}
