% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_318.R
\name{petplay}
\alias{petplay}
\title{How often r spends time each day playing with or training pet}
\source{
General Social Survey https://gss.norc.org
}
\description{
petplay
}
\details{
Question Thinking about your favorite pet, how often do you spend time each day playing with or exercising your pet? Would you say almost always, often, sometimes, or never?
}
\section{Values}{

\itemize{
\item \code{1} almost always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7207/vshow}{GSS Data Explorer page for \code{petplay}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab almost always \tab often \tab sometimes \tab never \tab no answer \tab Total \cr
   2018 \tab 1649 \tab 266 \tab 216 \tab 140 \tab 50 \tab 27 \tab 2348 \cr
   Total \tab 1649 \tab 266 \tab 216 \tab 140 \tab 50 \tab 27 \tab 2348 \cr
}
}

\keyword{variable}
