% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_311.R
\name{phyeffrt}
\alias{phyeffrt}
\title{Rate physical effort}
\source{
General Social Survey https://gss.norc.org
}
\description{
phyeffrt
}
\details{
Question A. Please rate the overall physical effort at the job you normally Do.
}
\section{Values}{

\itemize{
\item \verb{[1]} very hard
\item \verb{[2]} hard
\item \verb{[3]} somewhat hard
\item \verb{[4]} fairly light
\item \verb{[5]} very light
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2836/vshow}{GSS Data Explorer page for \code{phyeffrt}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very hard \tab hard \tab somewhat hard \tab fairly light \tab very light \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2010 \tab 857 \tab 101 \tab 133 \tab 294 \tab 329 \tab 302 \tab 3 \tab 25 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 89 \tab 164 \tab 338 \tab 344 \tab 306 \tab 2 \tab 7 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 105 \tab 154 \tab 431 \tab 333 \tab 390 \tab 5 \tab 30 \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 153 \tab 215 \tab 514 \tab 568 \tab 507 \tab 2 \tab 13 \tab 17 \tab 3544 \cr
   Total \tab 4600 \tab 448 \tab 666 \tab 1577 \tab 1574 \tab 1505 \tab 12 \tab 75 \tab 17 \tab 10474 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{backpain}},
\code{\link{hlthdays}},
\code{\link{hurtatwk}},
\code{\link{mntlhlth}},
\code{\link{painarms}},
\code{\link{physhlth}},
\code{\link{spvtrfair}},
\code{\link{strredpg}},
\code{\link{usedup}}
}
\concept{Quality of Working Life}
\keyword{variable}
