% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_651.R
\name{physact}
\alias{physact}
\title{How often r does physical activity for 20 min a day}
\source{
General Social Survey https://gss.norc.org
}
\description{
physact
}
\details{
Question (How often do you...) do physical activity for at least 20 minutes that makes you sweat or breathe more heavily than usual?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} once a month or less often
\item \code{3} several times a month
\item \code{4} several times a week
\item \code{5} daily
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7836/vshow}{GSS Data Explorer page for \code{physact}}.
}

\keyword{variable}
