% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_315.R
\name{physacts}
\alias{physacts}
\title{To what extent r can accomplish everyday physical activities}
\source{
General Social Survey https://gss.norc.org
}
\description{
physacts
}
\details{
Question To what extent are you able to carry out your everyday physical activities such as walking, climbing stairs, carrying groceries, or moving a chair?
}
\section{Values}{

\itemize{
\item \code{1} completely
\item \code{2} mostly
\item \code{3} moderately
\item \code{4} a little
\item \code{5} not at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7168/vshow}{GSS Data Explorer page for \code{physacts}}.
}

\keyword{variable}
