% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_311.R
\name{physhlth}
\alias{physhlth}
\title{Days of poor physical health past 30 days}
\source{
General Social Survey https://gss.norc.org
}
\description{
physhlth
}
\details{
Question A. Now thinking about your physical health, which includes physical illness and injury, for how many days during the past 30 days was your physical health not good?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2827/vshow}{GSS Data Explorer page for \code{physhlth}}.
}

\seealso{
Other Quality of Working Life: 
\code{\link{backpain}},
\code{\link{hlthdays}},
\code{\link{hurtatwk}},
\code{\link{mntlhlth}},
\code{\link{painarms}},
\code{\link{phyeffrt}},
\code{\link{spvtrfair}},
\code{\link{strredpg}},
\code{\link{usedup}}

Other Health: 
\code{\link{backpain}},
\code{\link{hlthdays}},
\code{\link{painarms}},
\code{\link{usedup}}
}
\concept{Health}
\concept{Quality of Working Life}
\keyword{variable}
