% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_552.R
\name{pikupsex}
\alias{pikupsex}
\title{R had sex with casual date last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
pikupsex
}
\details{
Question If you had other partners, please indicate all categories that apply to them.
C. Casual date or pick-up
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014 \cr
   B/C/- \tab 2016, 2018, 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5053/vshow}{GSS Data Explorer page for \code{pikupsex}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   1988 \tab 1243 \tab 48 \tab 122 \tab 68 \tab - \tab 1481 \cr
   1989 \tab 1303 \tab 42 \tab 118 \tab 74 \tab - \tab 1537 \cr
   1990 \tab 1155 \tab 45 \tab 101 \tab 71 \tab - \tab 1372 \cr
   1991 \tab 1293 \tab 47 \tab 110 \tab 67 \tab - \tab 1517 \cr
   1993 \tab 1359 \tab 58 \tab 102 \tab 87 \tab - \tab 1606 \cr
   1994 \tab 2489 \tab 100 \tab 234 \tab 169 \tab - \tab 2992 \cr
   1996 \tab 2401 \tab 109 \tab 275 \tab 119 \tab - \tab 2904 \cr
   1998 \tab 2383 \tab 91 \tab 204 \tab 154 \tab - \tab 2832 \cr
   2000 \tab 2335 \tab 104 \tab 201 \tab 177 \tab - \tab 2817 \cr
   2002 \tab 2317 \tab 137 \tab 264 \tab 46 \tab 1 \tab 2765 \cr
   2004 \tab 2429 \tab 136 \tab 211 \tab 36 \tab - \tab 2812 \cr
   2006 \tab 4087 \tab 147 \tab 260 \tab 16 \tab - \tab 4510 \cr
   2008 \tab 1688 \tab 118 \tab 194 \tab 23 \tab - \tab 2023 \cr
   2010 \tab 1705 \tab 116 \tab 193 \tab 30 \tab - \tab 2044 \cr
   2012 \tab 1669 \tab 105 \tab 165 \tab 35 \tab - \tab 1974 \cr
   2014 \tab 2144 \tab 143 \tab 199 \tab 52 \tab - \tab 2538 \cr
   2016 \tab 2530 \tab 141 \tab 176 \tab 20 \tab - \tab 2867 \cr
   2018 \tab 2130 \tab 78 \tab 117 \tab 21 \tab 2 \tab 2348 \cr
   2021 \tab 3746 \tab 125 \tab 130 \tab 20 \tab 11 \tab 4032 \cr
   2022 \tab 3304 \tab 98 \tab 83 \tab 41 \tab 18 \tab 3544 \cr
   Total \tab 43710 \tab 1988 \tab 3459 \tab 1326 \tab 32 \tab 50515 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{acqntsex}},
\code{\link{frndsex}},
\code{\link{matesex}},
\code{\link{othersex}},
\code{\link{paidsex}},
\code{\link{partners}},
\code{\link{sexfreq}},
\code{\link{sexsex}}
}
\concept{High Risk Behaviors}
\keyword{variable}
