% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_69.R
\name{pill}
\alias{pill}
\title{Birth control information}
\source{
General Social Survey https://gss.norc.org
}
\description{
pill
}
\details{
Question A. In some places in the United States, it is not legal to supply birth control information. How do you feel about this--do you think birth control information should be available to anyone who wants it, or not?
}
\section{Values}{

\itemize{
\item \verb{[1]} available
\item \verb{[2]} not available
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1974, 1975, 1977, 1982, 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/623/vshow}{GSS Data Explorer page for \code{pill}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab available \tab not available \tab don't know \tab no answer \tab Total \cr
   1974 \tab - \tab 1354 \tab 115 \tab 15 \tab - \tab 1484 \cr
   1975 \tab - \tab 1331 \tab 126 \tab 31 \tab 2 \tab 1490 \cr
   1977 \tab - \tab 1389 \tab 112 \tab 24 \tab 5 \tab 1530 \cr
   1982 \tab - \tab 1664 \tab 148 \tab 45 \tab 3 \tab 1860 \cr
   1983 \tab - \tab 1441 \tab 127 \tab 29 \tab 2 \tab 1599 \cr
   Total \tab 0 \tab 7179 \tab 628 \tab 144 \tab 12 \tab 7963 \cr
}
}

\seealso{
Other Core: 
\code{\link{abcare}},
\code{\link{abfirm}},
\code{\link{chldidel}},
\code{\link{chldmore}},
\code{\link{chldnum}},
\code{\link{chldsoon}},
\code{\link{pillok}},
\code{\link{sexeduc}},
\code{\link{teenpill}}
}
\concept{Core}
\keyword{variable}
