% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_69.R
\name{pillok}
\alias{pillok}
\title{Birth control to teenagers 14-16}
\source{
General Social Survey https://gss.norc.org
}
\description{
pillok
}
\details{
Question C. Do you strongly agree, agree, disagree, or strongly disagree that methods of birth control should be available to teenagers between the ages of 14 and 16 if their parents do not approve?
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1986 \tab -/- \tab full \cr
   1988 \tab A/B \tab partial \cr
   1989 \tab A/B \tab partial \cr
   1990 \tab A/B \tab partial \cr
   1991 \tab A/B \tab partial \cr
   1993 \tab A/B \tab partial \cr
   1994 \tab A/B \tab partial \cr
   1996 \tab A/B \tab partial \cr
   1998 \tab A/B \tab partial \cr
   2000 \tab A/B \tab partial \cr
   2002 \tab A/B \tab partial \cr
   2004 \tab A/B \tab partial \cr
   2006 \tab A/B \tab partial \cr
   2008 \tab A/B \tab partial \cr
   2010 \tab A/B \tab partial \cr
   2012 \tab A/B \tab partial \cr
   2014 \tab A/B \tab partial \cr
   2016 \tab A/B \tab partial \cr
   2018 \tab A/B \tab partial \cr
   2021 \tab A/B \tab partial \cr
   2022 \tab A/B \tab partial \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/625/vshow}{GSS Data Explorer page for \code{pillok}}.
}

\seealso{
Other Core: 
\code{\link{abcare}},
\code{\link{abfirm}},
\code{\link{chldidel}},
\code{\link{chldmore}},
\code{\link{chldnum}},
\code{\link{chldsoon}},
\code{\link{pill}},
\code{\link{sexeduc}},
\code{\link{teenpill}}
}
\concept{Core}
\keyword{variable}
