% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_607.R
\name{pilloky}
\alias{pilloky}
\title{Birth control to teenagers 14-16 (form 2)}
\source{
General Social Survey https://gss.norc.org
}
\description{
pilloky
}
\details{
Question Do strongly disagree, disagree, agree, or strongly agree that methods of birth control should be available to teenagers between the ages of 14 and 16 if their parents do not approve?
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5570/vshow}{GSS Data Explorer page for \code{pilloky}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1342 \tab 146 \tab 216 \tab 166 \tab 140 \tab 12 \tab 1 \tab 2023 \cr
   Total \tab 1342 \tab 146 \tab 216 \tab 166 \tab 140 \tab 12 \tab 1 \tab 2023 \cr
}
}

\concept{Form Y}
\keyword{variable}
