% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_503.R
\name{placewrk}
\alias{placewrk}
\title{Where do you work}
\source{
General Social Survey https://gss.norc.org
}
\description{
placewrk
}
\details{
Question Which of the following statements best describes where you work?
}
\section{Values}{

\itemize{
\item \code{1} in 1 place away
\item \code{2} at home
\item \code{3} part home/ away
\item \code{4} variety places
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4580/vshow}{GSS Data Explorer page for \code{placewrk}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab in 1 place away \tab at home \tab part home/ away \tab variety places \tab no answer \tab Total \cr
   1998 \tab 1967 \tab 616 \tab 31 \tab 38 \tab 146 \tab 34 \tab 2832 \cr
   Total \tab 1967 \tab 616 \tab 31 \tab 38 \tab 146 \tab 34 \tab 2832 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paytime}},
\code{\link{sethours}},
\code{\link{techjobs}},
\code{\link{techwork}},
\code{\link{wantjob1}},
\code{\link{worknow}}
}
\concept{Work Orientation}
\keyword{variable}
