% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_629.R
\name{planetrp}
\alias{planetrp}
\title{Plane trip in the past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
planetrp
}
\details{
Question In the last twelve months, how many trips did you make by plane? Count outward and return journeys, including transfers, as one trip.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7497/vshow}{GSS Data Explorer page for \code{planetrp}}.
}

\keyword{variable}
