% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_501.R
\name{planwrk}
\alias{planwrk}
\title{How much can r plan daily work}
\source{
General Social Survey https://gss.norc.org
}
\description{
planwrk
}
\details{
Question And which of the following statements about your work is most true?
}
\section{Values}{

\itemize{
\item \verb{[1]} plans most
\item \verb{[2]} plans part
\item \verb{[3]} cant plan
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4557/vshow}{GSS Data Explorer page for \code{planwrk}}.

Counts by year:\tabular{llllllll}{
   year \tab not available in this year \tab plans most \tab plans part \tab cant plan \tab no answer \tab iap \tab Total \cr
   1989 \tab - \tab 384 \tab 312 \tab 176 \tab 665 \tab - \tab 1537 \cr
   Total \tab 0 \tab 384 \tab 312 \tab 176 \tab 665 \tab 0 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{boredom}},
\code{\link{bossemps}},
\code{\link{cowrkers}},
\code{\link{danger}},
\code{\link{findwork}},
\code{\link{jobsat}},
\code{\link{stress}},
\code{\link{unhlthy}},
\code{\link{unpleznt}}
}
\concept{Work Orientation}
\keyword{variable}
