% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_268.R
\name{polagree}
\alias{polagree}
\title{Visit political sites r agrees with}
\source{
General Social Survey https://gss.norc.org
}
\description{
polagree
}
\details{
Question Is it VERY TRUE for you, SOMEWHAT TRUE, or NOT TRUE that you:
A. Visit political sites with points of view that you agree with?
}
\section{Values}{

\itemize{
\item \verb{[1]} very true
\item \verb{[2]} somewhat true
\item \verb{[3]} not true
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2439/vshow}{GSS Data Explorer page for \code{polagree}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not true \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2453 \tab 42 \tab 103 \tab 124 \tab 10 \tab 85 \tab 2817 \cr
   2002 \tab 2411 \tab 62 \tab 108 \tab 138 \tab 1 \tab 45 \tab 2765 \cr
   Total \tab 4864 \tab 104 \tab 211 \tab 262 \tab 11 \tab 130 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{enviro12}},
\code{\link{fe12}},
\code{\link{foraff12}},
\code{\link{gun12}},
\code{\link{moral12}},
\code{\link{news12}},
\code{\link{polcam12}},
\code{\link{racrel12}},
\code{\link{taxes12}}
}
\concept{Information Society}
\keyword{variable}
