% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_270.R
\name{polalter}
\alias{polalter}
\title{Web help change mind}
\source{
General Social Survey https://gss.norc.org
}
\description{
polalter
}
\details{
Question In general, when you visit Web sites devoted to political or social policy issues, current affairs, or political campaigns, is it VERY TRUE for you, SOMEWHAT TRUE, or NOT TRUE that:
C. Visiting such sites leads you to change your mind on issues?
}
\section{Values}{

\itemize{
\item \code{1} very true
\item \code{2} somewhat true
\item \code{3} not true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2449/vshow}{GSS Data Explorer page for \code{polalter}}.
}

\seealso{
Other Information Society: 
\code{\link{artschat}},
\code{\link{artschng}},
\code{\link{artshr}},
\code{\link{artsmin}},
\code{\link{artsnew}},
\code{\link{musicget}},
\code{\link{musicinf}},
\code{\link{poltough}},
\code{\link{wwwmusic}}
}
\concept{Information Society}
\keyword{variable}
