% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_269.R
\name{polchnge}
\alias{polchnge}
\title{Web change r's mind}
\source{
General Social Survey https://gss.norc.org
}
\description{
polchnge
}
\details{
Question Now I would like you to think about the last time you used the Web to get information or exchange views about a political or social-policy issue, current affairs, or a political campaign.
B. Did your Web experience change your view in any way?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2446/vshow}{GSS Data Explorer page for \code{polchnge}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2484 \tab 54 \tab 184 \tab 1 \tab 94 \tab 2817 \cr
   2002 \tab 2414 \tab 90 \tab 215 \tab 1 \tab 45 \tab 2765 \cr
   Total \tab 4898 \tab 144 \tab 399 \tab 2 \tab 139 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{contact}},
\code{\link{meeting}},
\code{\link{petition}},
\code{\link{polconf}},
\code{\link{poldsagr}},
\code{\link{polforms}},
\code{\link{polneutl}},
\code{\link{polnew}},
\code{\link{votefor}}
}
\concept{Information Society}
\keyword{variable}
