% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_269.R
\name{poldsagr}
\alias{poldsagr}
\title{Visit political sites r disagrees with}
\source{
General Social Survey https://gss.norc.org
}
\description{
poldsagr
}
\details{
Question Is it VERY TRUE for you, SOMEWHAT TRUE, or NOT TRUE that you:
C. Visit sites that challenge your own opinions and provide different perspectives on political affairs?
}
\section{Values}{

\itemize{
\item \code{1} very true
\item \code{2} somewhat true
\item \code{3} not true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2441/vshow}{GSS Data Explorer page for \code{poldsagr}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not true \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2453 \tab 70 \tab 93 \tab 108 \tab 8 \tab 85 \tab 2817 \cr
   2002 \tab 2411 \tab 80 \tab 114 \tab 114 \tab 1 \tab 45 \tab 2765 \cr
   Total \tab 4864 \tab 150 \tab 207 \tab 222 \tab 9 \tab 130 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{contact}},
\code{\link{meeting}},
\code{\link{petition}},
\code{\link{polchnge}},
\code{\link{polconf}},
\code{\link{polforms}},
\code{\link{polneutl}},
\code{\link{polnew}},
\code{\link{votefor}}
}
\concept{Information Society}
\keyword{variable}
