% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_446.R
\name{poleff13}
\alias{poleff13}
\title{Have a pretty good understanding of issues}
\source{
General Social Survey https://gss.norc.org
}
\description{
poleff13
}
\details{
Question How much do you agree or disagree with each of the following statements? PLEASE CHECK ONE BOX ON EACH LINE.
C. I feel that I have a pretty good understanding of the important political issues facing our country
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1996 \cr
   A/C/- \tab 2012 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4097/vshow}{GSS Data Explorer page for \code{poleff13}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1572 \tab 103 \tab 596 \tab 286 \tab 233 \tab 41 \tab 45 \tab 28 \tab 2904 \cr
   2006 \tab 2992 \tab 280 \tab 633 \tab 281 \tab 236 \tab 77 \tab 11 \tab - \tab 4510 \cr
   2012 \tab 1312 \tab 96 \tab 303 \tab 112 \tab 102 \tab 22 \tab 24 \tab 3 \tab 1974 \cr
   2016 \tab 1477 \tab 182 \tab 667 \tab 310 \tab 176 \tab 34 \tab 7 \tab 14 \tab 2867 \cr
   Total \tab 7353 \tab 661 \tab 2199 \tab 989 \tab 747 \tab 174 \tab 87 \tab 45 \tab 12255 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gvinflu1}},
\code{\link{gvinflu2}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff14}},
\code{\link{poleff15}},
\code{\link{poleff16}},
\code{\link{polgbeco}},
\code{\link{prvdschl}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
