% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_446.R
\name{poleff14}
\alias{poleff14}
\title{Elections a good way of making gov pay attention}
\source{
General Social Survey https://gss.norc.org
}
\description{
poleff14
}
\details{
Question How much do you agree or disagree with each of the following statements? PLEASE CHECK ONE BOX ON EACH LINE.
D. Elections are a good way of making governments pay attention to what the people think
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4098/vshow}{GSS Data Explorer page for \code{poleff14}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1572 \tab 267 \tab 698 \tab 155 \tab 131 \tab 26 \tab 26 \tab 29 \tab 2904 \cr
   Total \tab 1572 \tab 267 \tab 698 \tab 155 \tab 131 \tab 26 \tab 26 \tab 29 \tab 2904 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gvinflu1}},
\code{\link{gvinflu2}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff13}},
\code{\link{poleff15}},
\code{\link{poleff16}},
\code{\link{polgbeco}},
\code{\link{prvdschl}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
