% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_446.R
\name{poleff15}
\alias{poleff15}
\title{Most people better informed about politics than i}
\source{
General Social Survey https://gss.norc.org
}
\description{
poleff15
}
\details{
Question How much do you agree or disagree with each of the following statements? PLEASE CHECK ONE BOX ON EACH LINE.
E. I think most people are better informed about politics and government than I am
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
   A/C/- \tab 2012 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4099/vshow}{GSS Data Explorer page for \code{poleff15}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1572 \tab 92 \tab 323 \tab 348 \tab 423 \tab 88 \tab 34 \tab 24 \tab 2904 \cr
   2006 \tab 2992 \tab 158 \tab 420 \tab 280 \tab 512 \tab 137 \tab 11 \tab - \tab 4510 \cr
   2012 \tab 1312 \tab 37 \tab 197 \tab 143 \tab 218 \tab 53 \tab 11 \tab 3 \tab 1974 \cr
   Total \tab 5876 \tab 287 \tab 940 \tab 771 \tab 1153 \tab 278 \tab 56 \tab 27 \tab 9388 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gvinflu1}},
\code{\link{gvinflu2}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff13}},
\code{\link{poleff14}},
\code{\link{poleff16}},
\code{\link{polgbeco}},
\code{\link{prvdschl}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
