% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_446.R
\name{poleff16}
\alias{poleff16}
\title{People elected to congress try to keep promises}
\source{
General Social Survey https://gss.norc.org
}
\description{
poleff16
}
\details{
Question How much do you agree or disagree with each of the following statements? PLEASE CHECK ONE BOX ON EACH LINE.
F. People we elect to Congress try to keep the promises they have made during the election
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2012 \tab A/C/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4100/vshow}{GSS Data Explorer page for \code{poleff16}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gvinflu1}},
\code{\link{gvinflu2}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff13}},
\code{\link{poleff14}},
\code{\link{poleff15}},
\code{\link{polgbeco}},
\code{\link{prvdschl}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
