% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_447.R
\name{poleff17}
\alias{poleff17}
\title{Most government administrator's can be trusted}
\source{
General Social Survey https://gss.norc.org
}
\description{
poleff17
}
\details{
Question How much do you agree or disagree with each of the following statements? PLEASE CHECK ONE BOX ON EACH LINE.
G. Most government administrators can be trusted to do what is best for the country
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1996 \cr
   A/C/- \tab 2012 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4101/vshow}{GSS Data Explorer page for \code{poleff17}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1572 \tab 39 \tab 214 \tab 326 \tab 492 \tab 192 \tab 43 \tab 26 \tab 2904 \cr
   2006 \tab 2992 \tab 50 \tab 394 \tab 309 \tab 508 \tab 233 \tab 24 \tab - \tab 4510 \cr
   2012 \tab 1312 \tab 9 \tab 120 \tab 176 \tab 230 \tab 102 \tab 21 \tab 4 \tab 1974 \cr
   2016 \tab 1477 \tab 22 \tab 270 \tab 394 \tab 512 \tab 168 \tab 12 \tab 12 \tab 2867 \cr
   Total \tab 7353 \tab 120 \tab 998 \tab 1205 \tab 1742 \tab 695 \tab 100 \tab 42 \tab 12255 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{brlawfl}},
\code{\link{brnotax}},
\code{\link{demworks}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{runpower}},
\code{\link{taxbylaw}},
\code{\link{taxpaid}},
\code{\link{taxspend}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
