% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_447.R
\name{poleff17}
\alias{poleff17}
\title{Most government administrator's can be trusted}
\source{
General Social Survey https://gss.norc.org
}
\description{
poleff17
}
\details{
Question How much do you agree or disagree with each of the following statements? PLEASE CHECK ONE BOX ON EACH LINE.
G. Most government administrators can be trusted to do what is best for the country
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2012 \tab A/C/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4101/vshow}{GSS Data Explorer page for \code{poleff17}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{brlawfl}},
\code{\link{brnotax}},
\code{\link{demworks}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{runpower}},
\code{\link{taxbylaw}},
\code{\link{taxpaid}},
\code{\link{taxspend}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
