% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_541.R
\name{poleff20}
\alias{poleff20}
\title{Most ppl are better informed about politics than r is}
\source{
General Social Survey https://gss.norc.org
}
\description{
poleff20
}
\details{
Question To what extent do you agree or disagree with the following statements?
C. I think most people in America are better informed about politics and government than I am.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4941/vshow}{GSS Data Explorer page for \code{poleff20}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 65 \tab 287 \tab 359 \tab 615 \tab 141 \tab 1 \tab 4 \tab 2812 \cr
   2014 \tab 1274 \tab 71 \tab 303 \tab 313 \tab 411 \tab 133 \tab 29 \tab 4 \tab 2538 \cr
   Total \tab 2614 \tab 136 \tab 590 \tab 672 \tab 1026 \tab 274 \tab 30 \tab 8 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actlaw}},
\code{\link{affctlaw}},
\code{\link{discpol}},
\code{\link{govdook}},
\code{\link{oppsegov}},
\code{\link{poleff18}},
\code{\link{poleff19}},
\code{\link{polgreed}},
\code{\link{polint1}}
}
\concept{ISSP Citizenship}
\keyword{variable}
