% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_440.R
\name{poleff3}
\alias{poleff3}
\title{Avg person can influence politicians}
\source{
General Social Survey https://gss.norc.org
}
\description{
poleff3
}
\details{
Question Please indicate whether you agree or disagree with each of the following statements.
C. The average citizen has considerable influence on politics.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree-strongly agree
\item \verb{[2]} disagree-agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
   A/C/- \tab 2012 \cr
   D/-/- \tab 2006 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4041/vshow}{GSS Data Explorer page for \code{poleff3}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree-strongly agree \tab disagree-agree \tab can't choose \tab no answer \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab Total \cr
   1985 \tab 857 \tab 204 \tab 419 \tab 42 \tab 12 \tab - \tab - \tab - \tab 1534 \cr
   1996 \tab 1572 \tab 53 \tab 354 \tab 39 \tab 28 \tab 248 \tab 475 \tab 135 \tab 2904 \cr
   2006 \tab 2992 \tab 105 \tab 406 \tab 14 \tab 1 \tab 243 \tab 521 \tab 228 \tab 4510 \cr
   2012 \tab 1312 \tab 43 \tab 153 \tab 24 \tab 3 \tab 121 \tab 210 \tab 108 \tab 1974 \cr
   Total \tab 6733 \tab 405 \tab 1332 \tab 119 \tab 44 \tab 612 \tab 1206 \tab 471 \tab 10922 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{mustret}},
\code{\link{poleff1}},
\code{\link{poleff2}},
\code{\link{poleff4}},
\code{\link{poleff5}},
\code{\link{poleff6}},
\code{\link{poleff7}},
\code{\link{poleff8}},
\code{\link{poleff9}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
