% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_433.R
\name{polefy16}
\alias{polefy16}
\title{How often congress keeps promises}
\source{
General Social Survey https://gss.norc.org
}
\description{
polefy16
}
\details{
Question How often do the people we elect to Congress try to keep the promises they have made during the election: Never, rarely, sometimes, very often, or extremely often?
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} rarely
\item \verb{[3]} sometimes
\item \verb{[4]} very often
\item \verb{[5]} extremely often
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3976/vshow}{GSS Data Explorer page for \code{polefy16}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab never \tab rarely \tab sometimes \tab very often \tab extremely often \tab don't know \tab no answer \tab Total \cr
   2012 \tab 1328 \tab 59 \tab 215 \tab 288 \tab 49 \tab 5 \tab 28 \tab 2 \tab 1974 \cr
   Total \tab 1328 \tab 59 \tab 215 \tab 288 \tab 49 \tab 5 \tab 28 \tab 2 \tab 1974 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{polefy11}},
\code{\link{polefy13}},
\code{\link{polefy15}},
\code{\link{polefy17}},
\code{\link{polefy3}},
\code{\link{prespop}},
\code{\link{volacty2}},
\code{\link{volactyr}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
