% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_269.R
\name{polforms}
\alias{polforms}
\title{Web help form opinions}
\source{
General Social Survey https://gss.norc.org
}
\description{
polforms
}
\details{
Question In general, when you visit Web sites devoted to political or social policy issues, current affairs, or political campaigns, is it VERY TRUE for you, SOMEWHAT TRUE, or NOT TRUE that:
A. Visiting such Web sites helps you form an opinion on issues about which you are undecided?
}
\section{Values}{

\itemize{
\item \code{1} very true
\item \code{2} somewhat true
\item \code{3} not true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2447/vshow}{GSS Data Explorer page for \code{polforms}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not true \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2453 \tab 42 \tab 152 \tab 72 \tab 9 \tab 89 \tab 2817 \cr
   2002 \tab 2411 \tab 47 \tab 168 \tab 90 \tab 3 \tab 46 \tab 2765 \cr
   Total \tab 4864 \tab 89 \tab 320 \tab 162 \tab 12 \tab 135 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{contact}},
\code{\link{meeting}},
\code{\link{petition}},
\code{\link{polchnge}},
\code{\link{polconf}},
\code{\link{poldsagr}},
\code{\link{polneutl}},
\code{\link{polnew}},
\code{\link{votefor}}
}
\concept{Information Society}
\keyword{variable}
