% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_541.R
\name{polgreed}
\alias{polgreed}
\title{Most politicians are only for what get out of politics}
\source{
General Social Survey https://gss.norc.org
}
\description{
polgreed
}
\details{
Question To what extent do you agree or disagree with the following statements?
B. Most politicians are in politics only for what they can get out of it personally
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010 \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4946/vshow}{GSS Data Explorer page for \code{polgreed}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 195 \tab 508 \tab 346 \tab 390 \tab 23 \tab 4 \tab 6 \tab 2812 \cr
   2010 \tab 614 \tab 204 \tab 512 \tab 314 \tab 324 \tab 18 \tab 52 \tab 6 \tab 2044 \cr
   2014 \tab 1274 \tab 205 \tab 513 \tab 294 \tab 204 \tab 12 \tab 34 \tab 2 \tab 2538 \cr
   Total \tab 3228 \tab 604 \tab 1533 \tab 954 \tab 918 \tab 53 \tab 90 \tab 14 \tab 7394 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actlaw}},
\code{\link{affctlaw}},
\code{\link{discpol}},
\code{\link{govdook}},
\code{\link{oppsegov}},
\code{\link{poleff18}},
\code{\link{poleff19}},
\code{\link{poleff20}},
\code{\link{polint1}}
}
\concept{ISSP Citizenship}
\keyword{variable}
