% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_624.R
\name{polhitoky}
\alias{polhitoky}
\title{Ever approve of police striking citizen y}
\source{
General Social Survey https://gss.norc.org
}
\description{
polhitoky
}
\details{
Question Are there any situations you can imagine in which you would approve of a police officer striking an adult male citizen?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab B/C \tab partial \cr
   2022 \tab B/C \tab partial \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7447/vshow}{GSS Data Explorer page for \code{polhitoky}}.
}

\keyword{variable}
