% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_267.R
\name{polinf12}
\alias{polinf12}
\title{R use web for political information in past 12}
\source{
General Social Survey https://gss.norc.org
}
\description{
polinf12
}
\details{
Question In the past 12 months, that is, from (INSERT CURRENT MONTH) 1999/2001 through (INSERT CURRENT MONTH) 2000/02 have you used the Web to do each of these things:
M. Go to a Web site to find information about political or social policy issues, current affairs, or political campaigns?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2427/vshow}{GSS Data Explorer page for \code{polinf12}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 225 \tab 442 \tab 75 \tab - \tab 2817 \cr
   2002 \tab 1681 \tab 356 \tab 633 \tab 94 \tab 1 \tab 2765 \cr
   Total \tab 3756 \tab 581 \tab 1075 \tab 169 \tab 1 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{abort12}},
\code{\link{chat12}},
\code{\link{econ12}},
\code{\link{game12}},
\code{\link{health12}},
\code{\link{locate12}},
\code{\link{newjob12}},
\code{\link{people12}},
\code{\link{travel12}}
}
\concept{Information Society}
\keyword{variable}
