% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_279.R
\name{polinfgn}
\alias{polinfgn}
\title{R sought political information past year}
\source{
General Social Survey https://gss.norc.org
}
\description{
polinfgn
}
\details{
Question In the last twelve months, that is, from (INSERT CURRENT MONTH) 1999/2001 through (INSERT CURRENT MONTH) 2000/2002"have you:
H. Looked for information about political or social policy issues, current affairs, or political campaigns?
}
\section{Values}{

\itemize{
\item \verb{[0]} not at all
\item \verb{[1]} 1 or 2 times
\item \verb{[2]} 3-5 times
\item \verb{[3]} 6 or more times
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2000 \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2539/vshow}{GSS Data Explorer page for \code{polinfgn}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab not at all \tab 1 or 2 times \tab 3-5 times \tab no answer \tab 6 or more times \tab don't know \tab Total \cr
   2000 \tab 1884 \tab 554 \tab 160 \tab 200 \tab 19 \tab - \tab - \tab 2817 \cr
   2002 \tab 1372 \tab 784 \tab 175 \tab 135 \tab 6 \tab 285 \tab 8 \tab 2765 \cr
   Total \tab 3256 \tab 1338 \tab 335 \tab 335 \tab 25 \tab 285 \tab 8 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{artgn}},
\code{\link{byemail}},
\code{\link{byphone}},
\code{\link{inperson}},
\code{\link{letters}},
\code{\link{meetings}},
\code{\link{numcntct}},
\code{\link{perfrmgn}},
\code{\link{polcangn}}
}
\concept{Information Society}
\keyword{variable}
