% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_445.R
\name{polint}
\alias{polint}
\title{Your interest in politics}
\source{
General Social Survey https://gss.norc.org
}
\description{
polint
}
\details{
Question How interested would you say you personally are in politics?
}
\section{Values}{

\itemize{
\item \verb{[1]} very interested
\item \verb{[2]} fairly interested
\item \verb{[3]} somewhat interested
\item \verb{[4]} not very interested
\item \verb{[5]} not at all interested
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2024 \cr
   A/B/C \tab 1990, 1996 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4089/vshow}{GSS Data Explorer page for \code{polint}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very interested \tab fairly interested \tab somewhat interested \tab not very interested \tab not at all interested \tab can't choose \tab no answer \tab Total \cr
   1990 \tab 155 \tab 147 \tab 330 \tab 408 \tab 200 \tab 77 \tab 19 \tab 36 \tab 1372 \cr
   1996 \tab 1572 \tab 190 \tab 326 \tab 398 \tab 241 \tab 107 \tab 24 \tab 46 \tab 2904 \cr
   2006 \tab 2992 \tab 278 \tab 361 \tab 427 \tab 234 \tab 216 \tab 2 \tab - \tab 4510 \cr
   2016 \tab 1477 \tab 249 \tab 313 \tab 397 \tab 257 \tab 161 \tab 1 \tab 12 \tab 2867 \cr
   Total \tab 6196 \tab 864 \tab 1330 \tab 1630 \tab 932 \tab 561 \tab 46 \tab 94 \tab 11653 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gendereq}},
\code{\link{grnlaws}},
\code{\link{prvdhlth}},
\code{\link{prvdold}},
\code{\link{rprtst1}},
\code{\link{rprtst15}},
\code{\link{rprtst3}},
\code{\link{rprtst35}},
\code{\link{unionsok}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
