% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_445.R
\name{polint}
\alias{polint}
\title{Your interest in politics}
\source{
General Social Survey https://gss.norc.org
}
\description{
polint
}
\details{
Question How interested would you say you personally are in politics?
}
\section{Values}{

\itemize{
\item \code{1} very interested
\item \code{2} fairly interested
\item \code{3} somewhat interested
\item \code{4} not very interested
\item \code{5} not at all interested
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2024 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4089/vshow}{GSS Data Explorer page for \code{polint}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gendereq}},
\code{\link{grnlaws}},
\code{\link{prvdhlth}},
\code{\link{prvdold}},
\code{\link{rprtst1}},
\code{\link{rprtst15}},
\code{\link{rprtst3}},
\code{\link{rprtst35}},
\code{\link{unionsok}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
