% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_541.R
\name{polint1}
\alias{polint1}
\title{How interested in politics}
\source{
General Social Survey https://gss.norc.org
}
\description{
polint1
}
\details{
Question How interested would you say you personally are in politics?
}
\section{Values}{

\itemize{
\item \code{1} very interested
\item \code{2} fairly interested
\item \code{3} not very interested
\item \code{4} not at all interested
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4944/vshow}{GSS Data Explorer page for \code{polint1}}.
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actlaw}},
\code{\link{affctlaw}},
\code{\link{discpol}},
\code{\link{govdook}},
\code{\link{oppsegov}},
\code{\link{poleff18}},
\code{\link{poleff19}},
\code{\link{poleff20}},
\code{\link{polgreed}}
}
\concept{ISSP Citizenship}
\keyword{variable}
