% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_543.R
\name{polinter}
\alias{polinter}
\title{Expressed political views on internet past year}
\source{
General Social Survey https://gss.norc.org
}
\description{
polinter
}
\details{
Question Here are some different forms of political and social action that people can take. Please indicate, for each one, whether you have done any of these things in the past year, whether you have done it in the more distant past, whether you have not done it but might do it, or have not done it and would never, under any circumstances, do it.
I. Expressed political views on the internet
}
\section{Values}{

\itemize{
\item \verb{[1]} have done it in the past year
\item \verb{[2]} have done it in the more distant past
\item \verb{[3]} have not done it but might do it
\item \verb{[4]} have not done it and would never do it
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4965/vshow}{GSS Data Explorer page for \code{polinter}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab have done it in the past year \tab have done it in the more distant past \tab have not done it but might do it \tab have not done it and would never do  it \tab can't choose \tab no answer \tab Total \cr
   2014 \tab 1274 \tab 179 \tab 96 \tab 280 \tab 686 \tab 21 \tab 2 \tab 2538 \cr
   Total \tab 1274 \tab 179 \tab 96 \tab 280 \tab 686 \tab 21 \tab 2 \tab 2538 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{corruptn}},
\code{\link{crimlose}},
\code{\link{dem10fut}},
\code{\link{dem10pst}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{fixmistk}},
\code{\link{gvtrghts}},
\code{\link{polnews}}
}
\concept{ISSP Citizenship}
\keyword{variable}
