% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_672.R
\name{polintrst}
\alias{polintrst}
\title{Interested in campaigns this year}
\source{
General Social Survey https://gss.norc.org
}
\description{
polintrst
}
\details{
Question Some people don't pay much attention to political campaigns. How about you? Would you say that you have been very much interested, somewhat interested or not much interested in the political campaigns so far this year?
}
\section{Values}{

\itemize{
\item \code{1} very much interested
\item \code{2} somewhat interested
\item \code{3} not much interested
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8158/vshow}{GSS Data Explorer page for \code{polintrst}}.
}

\seealso{
Other ANES: 
\code{\link{econstat}}
}
\concept{ANES}
\keyword{variable}
