% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_86.R
\name{pollgood}
\alias{pollgood}
\title{Do surveys serve good purpose}
\source{
General Social Survey https://gss.norc.org
}
\description{
pollgood
}
\details{
Question In general, do you feel that surveys usually serve a good purpose or do you feel that they are usually a waste of time and money?
}
\section{Values}{

\itemize{
\item \verb{[1]} good purpose
\item \verb{[2]} depends
\item \verb{[3]} waste of time and $
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/776/vshow}{GSS Data Explorer page for \code{pollgood}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab good purpose \tab depends \tab waste of time and $ \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 1306 \tab 153 \tab 263 \tab 127 \tab 11 \tab 1860 \cr
   Total \tab 0 \tab 1306 \tab 153 \tab 263 \tab 127 \tab 11 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{bored}},
\code{\link{civic}},
\code{\link{feserve}},
\code{\link{imppol}},
\code{\link{meserve}},
\code{\link{polltrue}},
\code{\link{privacy}},
\code{\link{rushed}},
\code{\link{stresswk}}
}
\concept{Core}
\keyword{variable}
