% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_86.R
\name{polltrue}
\alias{polltrue}
\title{Trust results of surveys}
\source{
General Social Survey https://gss.norc.org
}
\description{
polltrue
}
\details{
Question How often do you think you can trust the results of surveys, Do you think they are almost always right, right most of the time, only some of the time, or hardly ever right?
}
\section{Values}{

\itemize{
\item \code{1} almost always
\item \code{2} most of the time
\item \code{3} some of the time
\item \code{4} hardly ever
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/777/vshow}{GSS Data Explorer page for \code{polltrue}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab almost always \tab most of the time \tab some of the time \tab hardly ever \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 157 \tab 722 \tab 802 \tab 82 \tab 85 \tab 12 \tab 1860 \cr
   Total \tab 0 \tab 157 \tab 722 \tab 802 \tab 82 \tab 85 \tab 12 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{bored}},
\code{\link{civic}},
\code{\link{feserve}},
\code{\link{imppol}},
\code{\link{meserve}},
\code{\link{pollgood}},
\code{\link{privacy}},
\code{\link{rushed}},
\code{\link{stresswk}}
}
\concept{Core}
\keyword{variable}
