% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_269.R
\name{polnew}
\alias{polnew}
\title{R learn something new on the web}
\source{
General Social Survey https://gss.norc.org
}
\description{
polnew
}
\details{
Question Now I would like you to think about the last time you used the Web to get information or exchange views about a political or social-policy issue, current affairs, or a political campaign.
A. Thinking back to this occasion, did you learn anything new about the issue or campaign?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[8]} cant recall
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2445/vshow}{GSS Data Explorer page for \code{polnew}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab can't choose \tab no answer \tab Total \cr
   2000 \tab 2453 \tab 130 \tab 116 \tab 31 \tab 87 \tab 2817 \cr
   2002 \tab 2411 \tab 169 \tab 137 \tab 3 \tab 45 \tab 2765 \cr
   Total \tab 4864 \tab 299 \tab 253 \tab 34 \tab 132 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{contact}},
\code{\link{meeting}},
\code{\link{petition}},
\code{\link{polchnge}},
\code{\link{polconf}},
\code{\link{poldsagr}},
\code{\link{polforms}},
\code{\link{polneutl}},
\code{\link{votefor}}
}
\concept{Information Society}
\keyword{variable}
