% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_543.R
\name{polnews}
\alias{polnews}
\title{How often use media to get political news}
\source{
General Social Survey https://gss.norc.org
}
\description{
polnews
}
\details{
Question How often do you use the media, including television, newspapers, radio and the internet, to get political news or information?
}
\section{Values}{

\itemize{
\item \verb{[1]} several times a day
\item \verb{[2]} once a day
\item \verb{[3]} 5-6 days a week
\item \verb{[4]} 3-4 days a week
\item \verb{[5]} 1-2 days a week
\item \verb{[6]} less than 1 day a week
\item \verb{[7]} never
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4966/vshow}{GSS Data Explorer page for \code{polnews}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab several times a day \tab once a day \tab 5-6 days a week \tab 3-4 days a week \tab 1-2 days a week \tab less than 1 day a week \tab never \tab can't choose \tab no answer \tab Total \cr
   2014 \tab 1274 \tab 344 \tab 371 \tab 90 \tab 91 \tab 101 \tab 145 \tab 111 \tab 10 \tab 1 \tab 2538 \cr
   Total \tab 1274 \tab 344 \tab 371 \tab 90 \tab 91 \tab 101 \tab 145 \tab 111 \tab 10 \tab 1 \tab 2538 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{corruptn}},
\code{\link{crimlose}},
\code{\link{dem10fut}},
\code{\link{dem10pst}},
\code{\link{demrghts}},
\code{\link{demtoday}},
\code{\link{fixmistk}},
\code{\link{gvtrghts}},
\code{\link{polinter}}
}
\concept{ISSP Citizenship}
\keyword{variable}
