% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_676.R
\name{polnewsfrom}
\alias{polnewsfrom}
\title{Where r gets political news}
\source{
General Social Survey https://gss.norc.org
}
\description{
polnewsfrom
}
\details{
Question Where do you get most of your news about political matters? Please select the most important source
}
\section{Values}{

\itemize{
\item \code{1} online social media (for example facebook, instagram, tiktok, x (formerly twitter), youtube, snapchat)
\item \code{2} television (including tv apps or access via internet)
\item \code{3} newspapers (in print or access via internet)
\item \code{4} radio and podcasts (including internet radio or access via internet
\item \code{5} websites (including exclusively online news sites, news aggregators, portals, or search engines)
\item \code{6} other
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8159/vshow}{GSS Data Explorer page for \code{polnewsfrom}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{harmgood1}},
\code{\link{inthrss}},
\code{\link{intoppo}},
\code{\link{intpolview}},
\code{\link{intscam}},
\code{\link{intviews}},
\code{\link{nextgen1}},
\code{\link{smnews}},
\code{\link{techesy}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
