% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_676.R
\name{polnewsfrom}
\alias{polnewsfrom}
\title{Where r gets political news}
\source{
General Social Survey https://gss.norc.org
}
\description{
polnewsfrom
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} online social media (for example facebook, instagram, tiktok, x (formerly twitter), youtube, snapchat)
\item \code{2} television (including tv apps or access via internet)
\item \code{3} newspapers (in print or access via internet)
\item \code{4} radio and podcasts (including internet radio or access via internet
\item \code{5} websites (including exclusively online news sites, news aggregators, portals, or search engines)
\item \code{6} other
\item \code{NA(d)} can't choose
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
