% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_449.R
\name{polsfair}
\alias{polsfair}
\title{Opinion of fairness of public official s treatment}
\source{
General Social Survey https://gss.norc.org
}
\description{
polsfair
}
\details{
Question In your opinion, how often do public officials deal fairly with people like you?
}
\section{Values}{

\itemize{
\item \code{1} almost always
\item \code{2} often
\item \code{3} occasionally
\item \code{4} seldom
\item \code{5} almost never
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4121/vshow}{GSS Data Explorer page for \code{polsfair}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{exploit}},
\code{\link{fewtrsty}},
\code{\link{givinffor}},
\code{\link{givinfusa}},
\code{\link{hlpinflu}},
\code{\link{stoprndm}},
\code{\link{tapphone}},
\code{\link{wotrial}},
\code{\link{youinflu}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
