% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_510.R
\name{polsgod}
\alias{polsgod}
\title{Atheists shouldn't hold public office}
\source{
General Social Survey https://gss.norc.org
}
\description{
polsgod
}
\details{
Question How much do you agree or disagree with each of the following…
A. Politicians who do not believe in God are unfit for public office.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4645/vshow}{GSS Data Explorer page for \code{polsgod}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{afterlif}},
\code{\link{churhpow}},
\code{\link{clerggov}},
\code{\link{clergvte}},
\code{\link{conchurh}},
\code{\link{concourt}},
\code{\link{conschls}},
\code{\link{godchnge}},
\code{\link{religpub}}
}
\concept{ISSP Religion}
\keyword{variable}
