% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_680.R
\name{poltalk}
\alias{poltalk}
\title{Politicians talk too much and don't take action}
\source{
General Social Survey https://gss.norc.org
}
\description{
poltalk
}
\details{
Question To what extent do you agree or disagree with the following statements?… Politicians talk too much and take too little action.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8060/vshow}{GSS Data Explorer page for \code{poltalk}}.
}

\seealso{
Other ISSP Citizenship: 
\code{\link{leftrght1}},
\code{\link{servpub}},
\code{\link{usmca1}}
}
\concept{ISSP Citizenship}
\keyword{variable}
