% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_270.R
\name{poltough}
\alias{poltough}
\title{Web show issues more complicated}
\source{
General Social Survey https://gss.norc.org
}
\description{
poltough
}
\details{
Question In general, when you visit Web sites devoted to political or social policy issues, current affairs, or political campaigns, is it VERY TRUE for you, SOMEWHAT TRUE, or NOT TRUE that:
D. After visiting such sites, you believe that issues are more complicated than you had previously thought?
}
\section{Values}{

\itemize{
\item \code{1} very true
\item \code{2} somewhat true
\item \code{3} not true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2450/vshow}{GSS Data Explorer page for \code{poltough}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not true \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2453 \tab 52 \tab 106 \tab 108 \tab 9 \tab 89 \tab 2817 \cr
   2002 \tab 2411 \tab 66 \tab 146 \tab 91 \tab 4 \tab 47 \tab 2765 \cr
   Total \tab 4864 \tab 118 \tab 252 \tab 199 \tab 13 \tab 136 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{artschat}},
\code{\link{artschng}},
\code{\link{artshr}},
\code{\link{artsmin}},
\code{\link{artsnew}},
\code{\link{musicget}},
\code{\link{musicinf}},
\code{\link{polalter}},
\code{\link{wwwmusic}}
}
\concept{Information Society}
\keyword{variable}
