% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_23.R
\name{polviewx}
\alias{polviewx}
\title{Liberal or conservative-version x}
\source{
General Social Survey https://gss.norc.org
}
\description{
polviewx
}
\details{
Question C. In political matters, people talk of "the left" and "the right."  How would you place your views on this scale?
(SHOW CARD S. Do NOT PROMPT. IF RESPONDENT HESITATES, ASK R TO TRY AGAIN. MARK ANSWER BELOW.)
}
\section{Values}{

\itemize{
\item \code{1} left
\item \code{10} right
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1983 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/180/vshow}{GSS Data Explorer page for \code{polviewx}}.
}

\seealso{
Other Core: 
\code{\link{if08who}},
\code{\link{if12who}},
\code{\link{polviews}},
\code{\link{polviewy}},
\code{\link{pres12}},
\code{\link{vote12}}
}
\concept{Core}
\keyword{variable}
