% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_23.R
\name{polviewy}
\alias{polviewy}
\title{Liberal or conservative-version y}
\source{
General Social Survey https://gss.norc.org
}
\description{
polviewy
}
\details{
Question B. We hear a lot of talk these days about liberals and conservatives.
I'm going to show you a seven-point scale on which the political views that people might hold are arranged from extremely liberal--point 1--to extremely conservative--point 7. Where would you place yourself on this scale, or haven't you thought much about this?
}
\section{Values}{

\itemize{
\item \code{0} havent thght abt it
\item \code{1} extremely liberal
\item \code{2} liberal
\item \code{3} slightly liberal
\item \code{4} moderate
\item \code{5} slghtly conservative
\item \code{6} conservative
\item \code{7} extrmly conservative
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1978 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/179/vshow}{GSS Data Explorer page for \code{polviewy}}.
}

\seealso{
Other Core: 
\code{\link{if08who}},
\code{\link{if12who}},
\code{\link{polviews}},
\code{\link{polviewx}},
\code{\link{pres12}},
\code{\link{vote12}}
}
\concept{Core}
\keyword{variable}
