% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_314.R
\name{poorserv}
\alias{poorserv}
\title{R recieves poorer service in restaurants or stores}
\source{
General Social Survey https://gss.norc.org
}
\description{
poorserv
}
\details{
Question In your day-to-day life how often have any of the following things happened to you? You receive poorer service than other people at restaurants or stores.
}
\section{Values}{

\itemize{
\item \verb{[1]} almost every day
\item \verb{[2]} at least once a week
\item \verb{[3]} a few times a month
\item \verb{[4]} a few times a year
\item \verb{[5]} less than once a year
\item \verb{[6]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7159/vshow}{GSS Data Explorer page for \code{poorserv}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab almost every day \tab at least once a week \tab a few times a month \tab a few times a year \tab less than once a year \tab never \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2018 \tab 774 \tab 16 \tab 72 \tab 158 \tab 408 \tab 375 \tab 531 \tab 8 \tab 6 \tab - \tab 2348 \cr
   2021 \tab 1410 \tab 24 \tab 82 \tab 155 \tab 672 \tab 774 \tab 885 \tab 3 \tab - \tab 27 \tab 4032 \cr
   2022 \tab 1207 \tab 45 \tab 98 \tab 173 \tab 535 \tab 599 \tab 868 \tab 8 \tab - \tab 11 \tab 3544 \cr
   Total \tab 3391 \tab 85 \tab 252 \tab 486 \tab 1615 \tab 1748 \tab 2284 \tab 19 \tab 6 \tab 38 \tab 9924 \cr
}
}

\keyword{variable}
