% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_608.R
\name{popespky}
\alias{popespky}
\title{Pope is infallible on matter's of faith or morals (form 2)}
\source{
General Social Survey https://gss.norc.org
}
\description{
popespky
}
\details{
Question Under certain conditions, the pope is infallible when he speaks on matters of faith and morals.
}
\section{Values}{

\itemize{
\item \code{1} certainly false
\item \code{2} probably false
\item \code{3} i am uncertain whether this is false or true
\item \code{4} probably true
\item \code{5} certainly true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5571/vshow}{GSS Data Explorer page for \code{popespky}}.
}

\seealso{
Other Form Y: 
\code{\link{letdie1y}},
\code{\link{polhitoy}}
}
\concept{Form Y}
\concept{Religion}
\keyword{variable}
