% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_232.R
\name{popmusic}
\alias{popmusic}
\title{Went to a live performance of popular music}
\source{
General Social Survey https://gss.norc.org
}
\description{
popmusic
}
\details{
Question Next I'd like to ask about some leisure or recreational activities that people do during their free time. As I read each activity, can you tell me if it is something you have done in the past twelve months?
B. Go to a live performance of popular music like rock, country, or rap, not including school performances.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2130/vshow}{GSS Data Explorer page for \code{popmusic}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   1998 \tab 1387 \tab 551 \tab 884 \tab 10 \tab - \tab 2832 \cr
   2002 \tab 1393 \tab 580 \tab 778 \tab 6 \tab 8 \tab 2765 \cr
   Total \tab 2780 \tab 1131 \tab 1662 \tab 16 \tab 8 \tab 5597 \cr
}
}

\seealso{
Other Culture: 
\code{\link{drama}},
\code{\link{readfict}}

Other Religion and Culture: 
\code{\link{aimofart}},
\code{\link{drama}},
\code{\link{localart}},
\code{\link{natlart}},
\code{\link{readfict}},
\code{\link{relart}},
\code{\link{stateart}},
\code{\link{volarts}}
}
\concept{Culture}
\concept{Religion and Culture}
\keyword{variable}
