% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_71.R
\name{pornfirm}
\alias{pornfirm}
\title{How firm is r's opinion on porn issue}
\source{
General Social Survey https://gss.norc.org
}
\description{
pornfirm
}
\details{
Question C. How firm are you about (your opinion on/the way you feel about) pornography--would you way you are very likely to change (your opinion), somewhat likely to change, somewhat unlikely to change or very unlikely to change your opinion?
}
\section{Values}{

\itemize{
\item \verb{[1]} change very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} somewhat unlikely
\item \verb{[4]} change very unlikely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/645/vshow}{GSS Data Explorer page for \code{pornfirm}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab change very likely \tab somewhat likely \tab somewhat unlikely \tab change very unlikely \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 37 \tab 205 \tab 417 \tab 765 \tab 26 \tab 23 \tab 1473 \cr
   Total \tab 0 \tab 37 \tab 205 \tab 417 \tab 765 \tab 26 \tab 23 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{pornimp}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{spanking}},
\code{\link{xmovie}},
\code{\link{xmovie1}}
}
\concept{Core}
\keyword{variable}
