% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_71.R
\name{pornout}
\alias{pornout}
\title{Materials provide outlet}
\source{
General Social Survey https://gss.norc.org
}
\description{
pornout
}
\details{
Question The next questions are about pornography--books, movies, magazines, and photographs that show or describe sex activities.
I''m going to read some opinions about the effects of looking at or reading such sexual materials. As I read each one, please tell me if you think sexual materials do or do not have that effect.
READ EACH ITEM. CODE ONE FOR EACH
D. Sexual materials provide an outlet for bottled-up impulses
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/- \tab full \cr
   1975 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1978 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab B/C \tab partial \cr
   1989 \tab B/C \tab partial \cr
   1990 \tab B/C \tab partial \cr
   1991 \tab B/C \tab partial \cr
   1993 \tab B/C \tab partial \cr
   1994 \tab B/C \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/639/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab don't know \tab no \tab no answer \tab yes \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 151 \tab 528 \tab 5 \tab 820 \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 236 \tab 413 \tab 5 \tab 836 \tab - \tab 1490 \cr
   1976 \tab - \tab 235 \tab 427 \tab 8 \tab 829 \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 183 \tab 445 \tab 5 \tab 899 \tab - \tab 1532 \cr
   1980 \tab - \tab 204 \tab 393 \tab 6 \tab 865 \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 176 \tab 501 \tab 11 \tab 911 \tab - \tab 1599 \cr
   1984 \tab - \tab 177 \tab 399 \tab 10 \tab 887 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 145 \tab 447 \tab 8 \tab 870 \tab - \tab 1470 \cr
   1987 \tab - \tab 253 \tab 479 \tab 12 \tab 1075 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 132 \tab 303 \tab 5 \tab 557 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 141 \tab 271 \tab 6 \tab 617 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 129 \tab 239 \tab 3 \tab 528 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 139 \tab 307 \tab 4 \tab 567 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 138 \tab 275 \tab 10 \tab 634 \tab - \tab 1606 \cr
   1994 \tab 1965 \tab 132 \tab 303 \tab 5 \tab 587 \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 48960 \tab 2571 \tab 5730 \tab 103 \tab 11482 \tab 6853 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornrape}},
\code{\link{spanking}},
\code{\link{xmovie}},
\code{\link{xmovie1}}

Other Sex: 
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornrape}},
\code{\link{xmovie}},
\code{\link{xmovie1}}
}
\concept{Core}
\concept{Sex}
\keyword{variable}
