% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_452.R
\name{posslq}
\alias{posslq}
\title{Does r have marital partner}
\source{
General Social Survey https://gss.norc.org
}
\description{
posslq
}
\details{
Question Which of these statements applies to you?  CODE ONE ONLY.
A. I am married and living in the same household as my husband or wife.
B. I am living as married and my partner and I together live in the same household.
C. I have a husband or wife or steady partner but we don''t live in the same household.
D. I don''t have a steady partner.
}
\section{Values}{

\itemize{
\item \code{1} i am married and living in the same household as my husband or wife
\item \code{2} i am living as married and my partner and i together live in the same household
\item \code{3} i have a husband or wife or steady partner, but we don't live in the same household
\item \code{4} i don't have a steady partner
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1986 \tab -/-/- \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4151/vshow}{GSS Data Explorer page for \code{posslq}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{brocall}},
\code{\link{daucall}},
\code{\link{daunum}},
\code{\link{dautime}},
\code{\link{dauvisit}},
\code{\link{soncall}},
\code{\link{sonnum}},
\code{\link{sontime}},
\code{\link{sonvisit}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
