% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_453.R
\name{posslqy}
\alias{posslqy}
\title{Relationship status and cohabitation or not}
\source{
General Social Survey https://gss.norc.org
}
\description{
posslqy
}
\details{
Question Which of these statements applies to you?  CODE ONE ONLY.
}
\section{Values}{

\itemize{
\item \code{1} i am married and living in the same household as my husband or wife
\item \code{2} i have a steady partner, and we live in the same household
\item \code{3} i have a husband or wife or steady partner, but we don't live in the same household
\item \code{4} i don't have a steady partner
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4152/vshow}{GSS Data Explorer page for \code{posslqy}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab i am married and living in the same household as my husband or wife \tab i have a steady partner, and we live in the same household \tab i have a husband or wife or steady partner, but we don't live in the same household \tab i don't have a steady partner \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   2012 \tab 998 \tab 433 \tab 97 \tab 96 \tab 348 \tab 2 \tab - \tab - \tab 1974 \cr
   2014 \tab 1274 \tab 562 \tab 120 \tab 117 \tab 462 \tab 1 \tab 2 \tab - \tab 2538 \cr
   2016 \tab 1442 \tab 576 \tab 149 \tab 130 \tab 570 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 1189 \tab 467 \tab 128 \tab 103 \tab 460 \tab 1 \tab - \tab - \tab 2348 \cr
   2021 \tab 1987 \tab 990 \tab 213 \tab 153 \tab 684 \tab - \tab - \tab 5 \tab 4032 \cr
   2022 \tab 1780 \tab 693 \tab 179 \tab 143 \tab 733 \tab 8 \tab 3 \tab 5 \tab 3544 \cr
   Total \tab 8670 \tab 3721 \tab 886 \tab 742 \tab 3257 \tab 12 \tab 5 \tab 10 \tab 17303 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{inlawnum}},
\code{\link{marcohab}},
\code{\link{relmost}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}},
\code{\link{unaunum}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
