% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_36.R
\name{postlf5}
\alias{postlf5}
\title{Pale or shadowy form of life}
\source{
General Social Survey https://gss.norc.org
}
\description{
postlf5
}
\details{
Question Do you believe there is a life after death?
A. Of course, no one knows exactly what life after death would be like, but here are some ideas people have had. (HAND CARD AA) How likely do you feel each possibility is? Would you say very likely, somewhat likely, not too likely, or not likely at all? (CIRCLE ONE NUMBER BESIDE EACH IDEA)
5. A pale, shadowy form of life, hardly life at all.
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not too likely
\item \verb{[4]} not likely at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/305/vshow}{GSS Data Explorer page for \code{postlf5}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not too likely \tab not likely at all \tab don't know \tab no answer \tab Total \cr
   1983 \tab 389 \tab 61 \tab 151 \tab 327 \tab 580 \tab 81 \tab 10 \tab 1599 \cr
   1984 \tab 278 \tab 63 \tab 154 \tab 312 \tab 567 \tab 81 \tab 18 \tab 1473 \cr
   Total \tab 667 \tab 124 \tab 305 \tab 639 \tab 1147 \tab 162 \tab 28 \tab 3072 \cr
}
}

\seealso{
Other Core: 
\code{\link{postlf1}},
\code{\link{postlf10}},
\code{\link{postlf2}},
\code{\link{postlf3}},
\code{\link{postlf4}},
\code{\link{postlf6}},
\code{\link{postlf7}},
\code{\link{postlf8}},
\code{\link{postlf9}}
}
\concept{Core}
\keyword{variable}
