% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_640.R
\name{postlifenv}
\alias{postlifenv}
\title{Belief in life after death (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
postlifenv
}
\details{
Question Do you believe there is a life after death?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[3]} undecided (phone mode only: volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7618/vshow}{GSS Data Explorer page for \code{postlifenv}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab skipped on web \tab Total \cr
   2021 \tab 2065 \tab 1407 \tab 514 \tab 18 \tab 28 \tab 4032 \cr
   2022 \tab 2688 \tab 596 \tab 245 \tab - \tab 15 \tab 3544 \cr
   Total \tab 4753 \tab 2003 \tab 759 \tab 18 \tab 43 \tab 7576 \cr
}
}

\keyword{variable}
