% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_638.R
\name{postlifev}
\alias{postlifev}
\title{Belief in life after death (with volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
postlifev
}
\details{
Question Do you believe there is a life after death?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[3]} undecided (phone mode only: volunteered)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7598/vshow}{GSS Data Explorer page for \code{postlifev}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2007 \tab 1180 \tab 350 \tab 485 \tab 1 \tab 9 \tab 4032 \cr
   2022 \tab 2698 \tab 490 \tab 120 \tab 233 \tab - \tab 3 \tab 3544 \cr
   Total \tab 4705 \tab 1670 \tab 470 \tab 718 \tab 1 \tab 12 \tab 7576 \cr
}
}

\keyword{variable}
