% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_519.R
\name{postmat1}
\alias{postmat1}
\title{Highest priority for america}
\source{
General Social Survey https://gss.norc.org
}
\description{
postmat1
}
\details{
Question A. Looking at the list below, please check a box next to the one thing you think should be America''s highest priority, the most important thing it should do. America should…
}
\section{Values}{

\itemize{
\item \code{1} maintain order in the nation
\item \code{2} give people more say in government decisions
\item \code{3} fight rising prices
\item \code{4} protect freedom of speech
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2010 \tab A/B/- \tab partial \cr
   2021 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4719/vshow}{GSS Data Explorer page for \code{postmat1}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{anrights}},
\code{\link{grnecon}},
\code{\link{harmgood}},
\code{\link{harmsgrn}},
\code{\link{postmat2}},
\code{\link{privent}},
\code{\link{scifaith}},
\code{\link{scigrn}},
\code{\link{sciworse}}
}
\concept{ISSP Environment}
\keyword{variable}
