% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_132.R
\name{povcol}
\alias{povcol}
\title{Give college aid to poor}
\source{
General Social Survey https://gss.norc.org
}
\description{
povcol
}
\details{
Question Here are several things that the government in Washington might Do to deal with the problems of poverty and
unemployment. I would like you to tell me if you favor or oppose them. (READ "A") -Would you say that you strongly favor it, favor it, neither favor nor oppose it, oppose, or strongly oppose it?
C. Provide special college scholarships for children from economically disadvantaged backgrounds who maintaIn good grades.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly favor
\item \verb{[2]} favor
\item \verb{[3]} neither favor nor oppose
\item \verb{[4]} oppose
\item \verb{[5]} strongly oppose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1232/vshow}{GSS Data Explorer page for \code{povcol}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly favor \tab favor \tab neither favor nor oppose \tab oppose \tab strongly oppose \tab don't know \tab no answer \tab Total \cr
   1990 \tab 698 \tab 256 \tab 349 \tab 37 \tab 12 \tab 7 \tab 7 \tab 6 \tab 1372 \cr
   Total \tab 698 \tab 256 \tab 349 \tab 37 \tab 12 \tab 7 \tab 7 \tab 6 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{blkcol}},
\code{\link{blkschs}},
\code{\link{blkzone}},
\code{\link{influblk}},
\code{\link{influjew}},
\code{\link{influwht}},
\code{\link{povschs}},
\code{\link{povzone}},
\code{\link{racquota}}
}
\concept{Intergroup Relations}
\keyword{variable}
