% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_567.R
\name{povline}
\alias{povline}
\title{R's hh poverty status}
\source{
General Social Survey https://gss.norc.org
}
\description{
povline
}
\details{
Question A. Poverty Status of Household
}
\section{Values}{

\itemize{
\item \code{1} poor
\item \code{2} poor -- imputed
\item \code{3} not poor
\item \code{4} not poor -- imputed
\item \code{5} borderline
\item \code{6} borderline -- imputed
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/- \tab full \cr
   1973 \tab -/-/- \tab full \cr
   1974 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1976 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5196/vshow}{GSS Data Explorer page for \code{povline}}.
}

\seealso{
Other Income: 
\code{\link{coninc}},
\code{\link{conrinc}},
\code{\link{incdef}},
\code{\link{minfour}},
\code{\link{realinc}},
\code{\link{realrinc}}

Other Household: 
\code{\link{incdef}}

Other Income: 
\code{\link{coninc}},
\code{\link{conrinc}},
\code{\link{incdef}},
\code{\link{minfour}},
\code{\link{realinc}},
\code{\link{realrinc}}

Other Poverty: 
\code{\link{incdef}}
}
\concept{Household}
\concept{Income}
\concept{Poverty}
\keyword{variable}
