% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_542.R
\name{powerun}
\alias{powerun}
\title{How much power r think united nations have}
\source{
General Social Survey https://gss.norc.org
}
\description{
powerun
}
\details{
Question Thinking about the United Nations, which comes closest to your view?
}
\section{Values}{

\itemize{
\item \code{1} un has too much power
\item \code{2} un has about right amount of power
\item \code{3} un has too little power
\item \code{4} dont know what un is
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4949/vshow}{GSS Data Explorer page for \code{powerun}}.
}

\seealso{
Other ISSP Citizenship: 
\code{\link{chngeoth}},
\code{\link{choices}},
\code{\link{elecfair}},
\code{\link{elecvote}},
\code{\link{govngos}},
\code{\link{polactve}},
\code{\link{refrndms}},
\code{\link{servepeo}},
\code{\link{unrghts}}

Other Political: 
\code{\link{chngeoth}},
\code{\link{choices}},
\code{\link{elecfair}},
\code{\link{elecvote}},
\code{\link{govngos}},
\code{\link{polactve}},
\code{\link{refrndms}},
\code{\link{servepeo}},
\code{\link{unrghts}}
}
\concept{ISSP Citizenship}
\concept{Political}
\keyword{variable}
