% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_639.R
\name{prayernv}
\alias{prayernv}
\title{Bible prayer in public schools (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
prayernv
}
\details{
Question The United States Supreme Court has ruled that no state or local government may require the reading of the Lord's Prayer or Bible verses in public schools. What are your views on this--do you approve or disapprove of the court ruling?
}
\section{Values}{

\itemize{
\item \code{1} approve
\item \code{2} disapprove
\item \code{3} no opinion (phone mode only: volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7605/vshow}{GSS Data Explorer page for \code{prayernv}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab approve \tab disapprove \tab don't know \tab skipped on web \tab Total \cr
   2021 \tab 2689 \tab 872 \tab 464 \tab 5 \tab 2 \tab 4032 \cr
   2022 \tab 2944 \tab 388 \tab 211 \tab - \tab 1 \tab 3544 \cr
   Total \tab 5633 \tab 1260 \tab 675 \tab 5 \tab 3 \tab 7576 \cr
}
}

\keyword{variable}
