% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_42.R
\name{prayerx}
\alias{prayerx}
\title{Bible prayer in schools:version z}
\source{
General Social Survey https://gss.norc.org
}
\description{
prayerx
}
\details{
Question C. Which of these statements comes closest to your feelings about prayers in the public schools?
The Lord''s prayer or some Bible verse should be read daily.
There should be a moment for silent prayer or meditation daily.
No prayer or other religious observances should be held in the public schools.
Other (SPECIFY).
}
\section{Values}{

\itemize{
\item \code{1} prayer read daily
\item \code{2} silent meditation
\item \code{3} no prayers
\item \code{4} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1986 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/363/vshow}{GSS Data Explorer page for \code{prayerx}}.
}

\seealso{
Other Core: 
\code{\link{bible}},
\code{\link{bibley}},
\code{\link{coden}},
\code{\link{cofund}},
\code{\link{cojew}},
\code{\link{coother}},
\code{\link{prayer}},
\code{\link{prayery}},
\code{\link{tithing}}
}
\concept{Core}
\keyword{variable}
