% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_513.R
\name{prayfreq}
\alias{prayfreq}
\title{How often does r pray}
\source{
General Social Survey https://gss.norc.org
}
\description{
prayfreq
}
\details{
Question Now thinking about the present, about how often do you pray?
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} less than once a year
\item \verb{[3]} about once or twice a year
\item \verb{[4]} several times a year
\item \verb{[5]} about once a month
\item \verb{[6]} 2-3 times a month
\item \verb{[7]} nearly every week
\item \verb{[8]} every week
\item \verb{[9]} several times a week
\item \verb{[10]} once a day
\item \verb{[11]} several times a day
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1991, 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4676/vshow}{GSS Data Explorer page for \code{prayfreq}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab never \tab less than once a year \tab about once or twice a year \tab several times a year \tab about once a month \tab 2-3 times a month \tab nearly every week \tab every week \tab several times a week \tab once a day \tab several times a day \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 105 \tab 37 \tab 73 \tab 110 \tab 65 \tab 71 \tab 77 \tab 61 \tab 166 \tab 289 \tab 271 \tab 1 \tab 33 \tab 1517 \cr
   1998 \tab 1548 \tab 132 \tab 41 \tab 41 \tab 54 \tab 49 \tab 57 \tab 62 \tab 63 \tab 164 \tab 306 \tab 275 \tab - \tab 40 \tab 2832 \cr
   2008 \tab 658 \tab 138 \tab 28 \tab 25 \tab 68 \tab 38 \tab 63 \tab 27 \tab 66 \tab 149 \tab 356 \tab 391 \tab 6 \tab 10 \tab 2023 \cr
   2018 \tab 1173 \tab 168 \tab 11 \tab 28 \tab 35 \tab 28 \tab 30 \tab 31 \tab 83 \tab 730 \tab - \tab - \tab 20 \tab 11 \tab 2348 \cr
   Total \tab 3537 \tab 543 \tab 117 \tab 167 \tab 267 \tab 180 \tab 221 \tab 197 \tab 273 \tab 1209 \tab 951 \tab 937 \tab 27 \tab 94 \tab 8720 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{densp}},
\code{\link{feelrel}},
\code{\link{relactiv}},
\code{\link{relactiv1}},
\code{\link{schlpray}}
}
\concept{ISSP Religion}
\keyword{variable}
