% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_511.R
\name{predeter}
\alias{predeter}
\title{God decides the course of our lives}
\source{
General Social Survey https://gss.norc.org
}
\description{
predeter
}
\details{
Question Do you agree or disagree with the following. . .
E. The course of our lives is decided by God.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4661/vshow}{GSS Data Explorer page for \code{predeter}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 229 \tab 258 \tab 272 \tab 273 \tab 202 \tab 65 \tab 60 \tab 1517 \cr
   Total \tab 158 \tab 229 \tab 258 \tab 272 \tab 273 \tab 202 \tab 65 \tab 60 \tab 1517 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{bible1}},
\code{\link{devil}},
\code{\link{fatalism}},
\code{\link{godmeans}},
\code{\link{heaven}},
\code{\link{hell}},
\code{\link{miracles}},
\code{\link{nihilism}},
\code{\link{theism}}
}
\concept{ISSP Religion}
\keyword{variable}
