% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_22.R
\name{pres08}
\alias{pres08}
\title{Vote obama or mccain}
\source{
General Social Survey https://gss.norc.org
}
\description{
pres08
}
\details{
Question K. In 2008, you remember that Obama ran for President on the Democratic ticket against McCain for the Republicans. Do you remember for sure whether or not you voted in that election?
\enumerate{
\item IF VOTED:  Did you vote for Obama or McCain?
}
}
\section{Values}{

\itemize{
\item \verb{[1]} obama
\item \verb{[2]} mccain
\item \verb{[3]} other candidate (specify)
\item \verb{[4]} didn't vote
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2010, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/173/vshow}{GSS Data Explorer page for \code{pres08}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab obama \tab mccain \tab other candidate (specify) \tab didn't vote \tab don't know \tab no answer \tab Total \cr
   2010 \tab 650 \tab 805 \tab 497 \tab 41 \tab 4 \tab 2 \tab 45 \tab 2044 \cr
   2012 \tab 666 \tab 774 \tab 448 \tab 27 \tab 10 \tab 10 \tab 39 \tab 1974 \cr
   2014 \tab 856 \tab 972 \tab 601 \tab 27 \tab 6 \tab 21 \tab 55 \tab 2538 \cr
   Total \tab 2172 \tab 2551 \tab 1546 \tab 95 \tab 20 \tab 33 \tab 139 \tab 6556 \cr
}
}

\seealso{
Other Core: 
\code{\link{if00who}},
\code{\link{if04who}},
\code{\link{if96who}},
\code{\link{pres00}},
\code{\link{pres04}},
\code{\link{pres96}},
\code{\link{vote00}},
\code{\link{vote04}},
\code{\link{vote08}}
}
\concept{Core}
\keyword{variable}
