% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_23.R
\name{pres12}
\alias{pres12}
\title{Vote obama or romney}
\source{
General Social Survey https://gss.norc.org
}
\description{
pres12
}
\details{
Question L. In 2012, you remember that Obama ran for President on the Democratic ticket against Romney for the Republicans. Do you remember for sure whether or not you voted in that election?
\enumerate{
\item Did you vote for Obama or Romney?
}
}
\section{Values}{

\itemize{
\item \verb{[1]} obama
\item \verb{[2]} romney
\item \verb{[3]} other candidate (specify)
\item \verb{[4]} didn't vote for president
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014, 2016, 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/176/vshow}{GSS Data Explorer page for \code{pres12}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab obama \tab romney \tab other candidate (specify) \tab didn't vote for president \tab don't know \tab no answer \tab Total \cr
   2014 \tab 879 \tab 927 \tab 613 \tab 45 \tab 14 \tab 10 \tab 50 \tab 2538 \cr
   2016 \tab 1049 \tab 1082 \tab 591 \tab 42 \tab 15 \tab 21 \tab 67 \tab 2867 \cr
   2018 \tab 850 \tab 925 \tab 473 \tab 20 \tab 8 \tab 23 \tab 49 \tab 2348 \cr
   Total \tab 2778 \tab 2934 \tab 1677 \tab 107 \tab 37 \tab 54 \tab 166 \tab 7753 \cr
}
}

\seealso{
Other Core: 
\code{\link{if08who}},
\code{\link{if12who}},
\code{\link{polviews}},
\code{\link{polviewx}},
\code{\link{polviewy}},
\code{\link{vote12}}
}
\concept{Core}
\keyword{variable}
