% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_23.R
\name{pres16}
\alias{pres16}
\title{Vote clinton or trump}
\source{
General Social Survey https://gss.norc.org
}
\description{
pres16
}
\details{
Question Did you vote for Clinton or Trump?
}
\section{Values}{

\itemize{
\item \code{1} clinton
\item \code{2} trump
\item \code{3} other candidate
\item \code{4} didn't vote for president
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7090/vshow}{GSS Data Explorer page for \code{pres16}}.
}

\keyword{variable}
