% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_643.R
\name{pres20}
\alias{pres20}
\title{Voted trump or biden}
\source{
General Social Survey https://gss.norc.org
}
\description{
pres20
}
\details{
Question Did you vote for Joe Biden or Donald Trump?
}
\section{Values}{

\itemize{
\item \verb{[1]} biden
\item \verb{[2]} trump
\item \verb{[3]} other candidate
\item \verb{[4]} didn't vote for president
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7759/vshow}{GSS Data Explorer page for \code{pres20}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab biden \tab trump \tab other candidate \tab didn't vote for president \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 939 \tab 1417 \tab 883 \tab 82 \tab 14 \tab 38 \tab 158 \tab 13 \tab 3544 \cr
   Total \tab 939 \tab 1417 \tab 883 \tab 82 \tab 14 \tab 38 \tab 158 \tab 13 \tab 3544 \cr
}
}

\keyword{variable}
