% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_19.R
\name{pres68}
\alias{pres68}
\title{Vote for humphrey, nixon, or wallace}
\source{
General Social Survey https://gss.norc.org
}
\description{
pres68
}
\details{
Question A. Now in 1968, you remember that Humphrey ran for President on the Democratic ticket against Nixon for the Republicans, and Wallace as an Independent. Do you remember for sure whether or not you voted in that election?
\enumerate{
\item IF VOTED:  Did you vote for Humphrey, Nixon, or Wallace?
}
}
\section{Values}{

\itemize{
\item \code{1} humphrey
\item \code{2} nixon
\item \code{3} wallace
\item \code{4} other
\item \code{5} refused
\item \code{6} wldnt vt-relig
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab - \tab full \cr
   1973 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/143/vshow}{GSS Data Explorer page for \code{pres68}}.
}

\seealso{
Other Core: 
\code{\link{padotdat}},
\code{\link{padotged}},
\code{\link{padotpeo}},
\code{\link{padotpre}},
\code{\link{padotsvp}},
\code{\link{padotthn}},
\code{\link{partyid}},
\code{\link{spdotpre}},
\code{\link{vote68}}

Other Political: 
\code{\link{partyid}},
\code{\link{vote68}}
}
\concept{Core}
\concept{Political}
\keyword{variable}
