% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_20.R
\name{pres72}
\alias{pres72}
\title{Vote for mcgovern or nixon}
\source{
General Social Survey https://gss.norc.org
}
\description{
pres72
}
\details{
Question B. In 1972, you remember that McGovern ran for President on the Democratic ticket against Nixon for the Republicans. Do you remember for sure whether or not you voted in that election?
\enumerate{
\item IF VOTED:  Did you vote for McGovern or Nixon?
}
}
\section{Values}{

\itemize{
\item \verb{[1]} mcgovern
\item \verb{[2]} nixon
\item \verb{[3]} other
\item \verb{[4]} refused
\item \verb{[5]} wouldnt vote
\item \verb{[6]} wldnt vt-relig
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973, 1974, 1975, 1976, 1977 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/146/vshow}{GSS Data Explorer page for \code{pres72}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab mcgovern \tab nixon \tab other \tab refused \tab don't know \tab no answer \tab wouldnt vote \tab Total \cr
   1973 \tab 459 \tab 404 \tab 593 \tab 18 \tab 20 \tab 3 \tab 7 \tab - \tab 1504 \cr
   1974 \tab 470 \tab 379 \tab 560 \tab 38 \tab 16 \tab 12 \tab 9 \tab - \tab 1484 \cr
   1975 \tab 526 \tab 328 \tab 567 \tab 28 \tab - \tab 11 \tab 17 \tab 13 \tab 1490 \cr
   1976 \tab 549 \tab 344 \tab 544 \tab 32 \tab - \tab 20 \tab 7 \tab 3 \tab 1499 \cr
   1977 \tab 610 \tab 329 \tab 540 \tab 14 \tab - \tab 19 \tab 12 \tab 6 \tab 1530 \cr
   Total \tab 2614 \tab 1784 \tab 2804 \tab 130 \tab 36 \tab 65 \tab 52 \tab 22 \tab 7507 \cr
}
}

\seealso{
Other Core: 
\code{\link{if68who}},
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{if80who}},
\code{\link{pres76}},
\code{\link{pres80}},
\code{\link{vote72}},
\code{\link{vote76}},
\code{\link{vote80}}
}
\concept{Core}
\keyword{variable}
