% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_20.R
\name{pres76}
\alias{pres76}
\title{Vote for carter or ford}
\source{
General Social Survey https://gss.norc.org
}
\description{
pres76
}
\details{
Question C. In 1976, you remember that Carter ran for President on the Democratic ticket against Ford for the Republicans. Do you remember for sure whether or not you voted in that election?
\enumerate{
\item IF VOTED:  Did you vote for Carter or Ford?
}
}
\section{Values}{

\itemize{
\item \code{1} carter
\item \code{2} ford
\item \code{3} other
\item \code{4} refused
\item \code{5} no pres. vote
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1977 \tab - \tab full \cr
   1978 \tab - \tab full \cr
   1980 \tab - \tab full \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/149/vshow}{GSS Data Explorer page for \code{pres76}}.
}

\seealso{
Other Core: 
\code{\link{if68who}},
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{if80who}},
\code{\link{pres72}},
\code{\link{pres80}},
\code{\link{vote72}},
\code{\link{vote76}},
\code{\link{vote80}}
}
\concept{Core}
\keyword{variable}
