% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_21.R
\name{pres92}
\alias{pres92}
\title{Vote for clinton, bush, perot}
\source{
General Social Survey https://gss.norc.org
}
\description{
pres92
}
\details{
Question G. In 1992, you remember that Clinton ran for President on the Democratic ticket against Bush for the Republicans and Perot as an Independent.
Do you remember for sure whether or not you voted in that election?
\enumerate{
\item IF VOTED:
Did you vote for Clinton, Bush, or Perot?
}
}
\section{Values}{

\itemize{
\item \code{1} clinton
\item \code{2} bush
\item \code{3} perot
\item \code{4} other
\item \code{6} no pres. vote
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/161/vshow}{GSS Data Explorer page for \code{pres92}}.
}

\seealso{
Other Core: 
\code{\link{if84who}},
\code{\link{if88who}},
\code{\link{if92who}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote92}},
\code{\link{vote96}}
}
\concept{Core}
\keyword{variable}
