% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_15.R
\name{preteen}
\alias{preteen}
\title{Household members 6 thru 12 yrs old}
\source{
General Social Survey https://gss.norc.org
}
\description{
preteen
}
\details{
Question Household Size and Composition (see Appendix D:  Recodes, for information about these variables). B. NUMBER OF MEMBERS 6 TO 12 YEARS OLD:
As of 2024, PRETEEN has been recoded to only account for 0 to 2+ children ages 6 to 12 in the household. This has been retroactively applied to all previous years of PRETEEN. The original PRETEEN containing data from 1972 to 2022 has been renamed PRETEEN_7222.
}
\section{Values}{

\itemize{
\item \code{0} 0 household members between 6 to 12
\item \code{8} 8 or more
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/99/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllllllllll}{
   year \tab 0 household members between 6 to 12 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 8 or more \tab no answer \tab skipped on web \tab iap \tab not available in this release \tab Total \cr
   1972 \tab 1159 \tab 252 \tab 135 \tab 47 \tab 14 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1101 \tab 212 \tab 124 \tab 47 \tab 19 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1066 \tab 224 \tab 137 \tab 49 \tab 1 \tab 5 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1127 \tab 202 \tab 119 \tab 30 \tab 7 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1139 \tab 221 \tab 99 \tab 32 \tab 5 \tab - \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1154 \tab 219 \tab 105 \tab 33 \tab 8 \tab 1 \tab 1 \tab - \tab 9 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1159 \tab 220 \tab 109 \tab 21 \tab 6 \tab 3 \tab - \tab - \tab 14 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1131 \tab 209 \tab 97 \tab 17 \tab 2 \tab - \tab - \tab - \tab 12 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1470 \tab 259 \tab 97 \tab 22 \tab 6 \tab 1 \tab - \tab - \tab 5 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1235 \tab 228 \tab 105 \tab 22 \tab 1 \tab 2 \tab - \tab - \tab 6 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1188 \tab 169 \tab 93 \tab 16 \tab 2 \tab - \tab - \tab - \tab 5 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1238 \tab 184 \tab 94 \tab 10 \tab 1 \tab 1 \tab 1 \tab - \tab 5 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1172 \tab 184 \tab 79 \tab 21 \tab 8 \tab - \tab - \tab - \tab 6 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1438 \tab 237 \tab 122 \tab 15 \tab 4 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1167 \tab 196 \tab 78 \tab 25 \tab 3 \tab - \tab - \tab - \tab 12 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1208 \tab 193 \tab 89 \tab 21 \tab 6 \tab - \tab - \tab - \tab 20 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1102 \tab 157 \tab 86 \tab 15 \tab 2 \tab - \tab - \tab - \tab 10 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1198 \tab 184 \tab 82 \tab 20 \tab 3 \tab 1 \tab - \tab - \tab 29 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1294 \tab 192 \tab 96 \tab 18 \tab 2 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2423 \tab 369 \tab 162 \tab 31 \tab 4 \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2368 \tab 346 \tab 155 \tab 29 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2315 \tab 311 \tab 161 \tab 33 \tab 3 \tab 1 \tab 2 \tab - \tab 6 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2200 \tab 377 \tab 152 \tab 28 \tab 9 \tab 4 \tab - \tab - \tab 47 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2337 \tab 260 \tab 109 \tab 26 \tab 3 \tab - \tab - \tab - \tab 30 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2358 \tab 282 \tab 126 \tab 21 \tab 6 \tab - \tab - \tab - \tab 19 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 3748 \tab 483 \tab 199 \tab 38 \tab 2 \tab 2 \tab 1 \tab - \tab 37 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1673 \tab 219 \tab 78 \tab 19 \tab 5 \tab - \tab - \tab - \tab 29 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1739 \tab 186 \tab 84 \tab 23 \tab 3 \tab - \tab 1 \tab 1 \tab 7 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1636 \tab 210 \tab 90 \tab 16 \tab 1 \tab - \tab - \tab - \tab 21 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2220 \tab 195 \tab 93 \tab 24 \tab 2 \tab - \tab - \tab - \tab 4 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2497 \tab 229 \tab 86 \tab 23 \tab 6 \tab - \tab - \tab - \tab 26 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2047 \tab 183 \tab 72 \tab 12 \tab 5 \tab - \tab - \tab - \tab 29 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 3544 \tab 313 \tab 131 \tab 38 \tab 3 \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab 4032 \cr
   2022 \tab 1544 \tab 138 \tab 67 \tab 16 \tab 3 \tab - \tab - \tab - \tab 8 \tab 4 \tab 1764 \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 57395 \tab 8043 \tab 3711 \tab 858 \tab 161 \tab 31 \tab 9 \tab 2 \tab 409 \tab 7 \tab 1764 \tab 3309 \tab 75699 \cr
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/-/- \tab full \cr
   1973 \tab -/-/-/- \tab full \cr
   1974 \tab -/-/-/- \tab full \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
   2024 \tab A/B/C/- \tab full \cr
}
}

\seealso{
Other Core: 
\code{\link{adults}},
\code{\link{babies}},
\code{\link{earnrs}},
\code{\link{granborn}},
\code{\link{hompop}},
\code{\link{income}},
\code{\link{rincome}},
\code{\link{teens}},
\code{\link{unrelat}}

Other Household: 
\code{\link{adults}},
\code{\link{babies}},
\code{\link{earnrs}},
\code{\link{hompop}},
\code{\link{teens}},
\code{\link{unrelat}}
}
\concept{Core}
\concept{Household}
\keyword{variable}
