% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_666.R
\name{preteen_exp}
\alias{preteen_exp}
\title{Kids 7-12 in hh(experimental version)}
\source{
General Social Survey https://gss.norc.org
}
\description{
preteen_exp
}
\details{
Question How many individuals are there in this household between the ages of 7 and 12?
As of 2024, PRETEEN_EXP has been recoded to only account for 0 to 2+ children ages 7 to 12 in the household. This has been retroactively applied to all previous years of PRETEEN_EXP. The original PRETEEN_EXP containing data from in 2022 has been renamed PRETEEN_EXP_22.
}
\section{Values}{

\itemize{
\item \code{0} 0 household members between 6 to 12
\item \code{8} 8 or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8008/vshow}{GSS Data Explorer page for \code{preteen_exp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab 0 household members between 6 to 12 \tab 1 \tab 2 \tab 3 \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 1788 \tab 1459 \tab 196 \tab 65 \tab 15 \tab 9 \tab 12 \tab 3544 \cr
   Total \tab 1788 \tab 1459 \tab 196 \tab 65 \tab 15 \tab 9 \tab 12 \tab 3544 \cr
}
}

\seealso{
Other Core: 
\code{\link{teens_exp}}
}
\concept{Core}
\keyword{variable}
