% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_236.R
\name{prfmcost}
\alias{prfmcost}
\title{Did not go due to cost}
\source{
General Social Survey https://gss.norc.org
}
\description{
prfmcost
}
\details{
Question Thinking about the most recent performance you wanted to attend but did not, which of the following factors were important in your decision not to attend?     
Cost too much.
}
\section{Values}{

\itemize{
\item \verb{[1]} a reason
\item \verb{[2]} not a reason
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6249/vshow}{GSS Data Explorer page for \code{prfmcost}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab a reason \tab not a reason \tab Total \cr
   2016 \tab 2408 \tab 199 \tab 260 \tab 2867 \cr
   Total \tab 2408 \tab 199 \tab 260 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{artwhy3}},
\code{\link{artwhy4}},
\code{\link{artwhy5}},
\code{\link{artwhy6}},
\code{\link{artwhy7}},
\code{\link{artwhy8}},
\code{\link{artwhy9}},
\code{\link{nogo}},
\code{\link{nogo1}}
}
\concept{Arts and Culture}
\keyword{variable}
