% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_232.R
\name{prfmnce}
\alias{prfmnce}
\title{Did r go to a performance in last 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
prfmnce
}
\details{
Question PAY SPECIAL ATTENTION TO THE HANDCARD NUMBER INDICATED FOR THE QUESTIONS ASKED.  DEPENDING ON THE RESPONDENT'S ANSWERS, YOU WILL NEED TO REFER HIM/HER TO HANDCARD H1, H2, H3 OR H4.    
With the exception of elementary or high school performances, did you go to a live music, theater, or dance performance, during the last 12 months?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6209/vshow}{GSS Data Explorer page for \code{prfmnce}}.
}

\concept{Arts and Culture}
\keyword{variable}
