% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_234.R
\name{prfmwhy9}
\alias{prfmwhy9}
\title{Was there another major reason in decision to attend perf}
\source{
General Social Survey https://gss.norc.org
}
\description{
prfmwhy9
}
\details{
Question Was there some other major reason for your decision to attend that is not shown on the card?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6229/vshow}{GSS Data Explorer page for \code{prfmwhy9}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2016 \tab 2433 \tab 75 \tab 351 \tab 8 \tab 2867 \cr
   Total \tab 2433 \tab 75 \tab 351 \tab 8 \tab 2867 \cr
}
}

\seealso{
Other Arts and Culture: 
\code{\link{prfmfree}},
\code{\link{prfmwhy1}},
\code{\link{prfmwhy2}},
\code{\link{prfmwhy3}},
\code{\link{prfmwhy4}},
\code{\link{prfmwhy5}},
\code{\link{prfmwhy6}},
\code{\link{prfmwhy7}},
\code{\link{prfmwhy8}}
}
\concept{Arts and Culture}
\keyword{variable}
