% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_505.R
\name{priagncy}
\alias{priagncy}
\title{Registered at a private employment agency}
\source{
General Social Survey https://gss.norc.org
}
\description{
priagncy
}
\details{
Question Thinking about the last 12 months, have you done any of the following in order to find a job?
B. Registered at a private employment agency?
}
\section{Values}{

\itemize{
\item \code{1} no
\item \code{2} yes, once or twice
\item \code{3} yes, more than twice
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4600/vshow}{GSS Data Explorer page for \code{priagncy}}.
}

\seealso{
Other Work Orientation: 
\code{\link{evjob}},
\code{\link{getjob}},
\code{\link{leavejob}},
\code{\link{lookjob}},
\code{\link{pubagncy}},
\code{\link{wantjob2}},
\code{\link{whyjbend}},
\code{\link{worryjob}},
\code{\link{yrjobend}}
}
\concept{Work Orientation}
\keyword{variable}
