% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_443.R
\name{pricecon}
\alias{pricecon}
\title{Govts resp: keep prices under control}
\source{
General Social Survey https://gss.norc.org
}
\description{
pricecon
}
\details{
Question On the whole, do you think it should or should not be the government's responsibility to . . .
B. Keep prices under control.
}
\section{Values}{

\itemize{
\item \code{1} definitely should be
\item \code{2} probably should be
\item \code{3} probably should not be
\item \code{4} definitely should not be
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1990, 1996 \cr
   D/-/- \tab 2006 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4077/vshow}{GSS Data Explorer page for \code{pricecon}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should be \tab probably should be \tab probably should not be \tab definitely should not be \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 194 \tab 293 \tab 106 \tab 53 \tab 18 \tab 13 \tab 1534 \cr
   1990 \tab 155 \tab 288 \tab 577 \tab 190 \tab 80 \tab 38 \tab 44 \tab 1372 \cr
   1996 \tab 1572 \tab 306 \tab 552 \tab 242 \tab 144 \tab 34 \tab 54 \tab 2904 \cr
   2006 \tab 2992 \tab 664 \tab 495 \tab 221 \tab 125 \tab 12 \tab 1 \tab 4510 \cr
   2016 \tab 1477 \tab 440 \tab 544 \tab 247 \tab 125 \tab 24 \tab 10 \tab 2867 \cr
   Total \tab 7053 \tab 1892 \tab 2461 \tab 1006 \tab 527 \tab 126 \tab 122 \tab 13187 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{buspow}},
\code{\link{govtpow}},
\code{\link{jobsall}},
\code{\link{laborpow}},
\code{\link{ownautos}},
\code{\link{ownbanks}},
\code{\link{ownmass}},
\code{\link{ownpower}},
\code{\link{ownsteel}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
