% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_504.R
\name{prideorg}
\alias{prideorg}
\title{I am proud working for my firm}
\source{
General Social Survey https://gss.norc.org
}
\description{
prideorg
}
\details{
Question To what extent do you agree or disagree with each of the following statements?
B. I am proud to be working for my firm or organization.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4586/vshow}{GSS Data Explorer page for \code{prideorg}}.
}

\seealso{
Other Work Orientation: 
\code{\link{absent}},
\code{\link{chngwork}},
\code{\link{edcskill}},
\code{\link{helporg1}},
\code{\link{jobskill}},
\code{\link{proudwrk}},
\code{\link{stayorg3}},
\code{\link{useskill}},
\code{\link{wktenure}}

Other Work: 
\code{\link{absent}},
\code{\link{chngwork}},
\code{\link{edcskill}},
\code{\link{helporg1}},
\code{\link{jobskill}},
\code{\link{proudwrk}},
\code{\link{stayorg3}},
\code{\link{useskill}},
\code{\link{wktenure}}
}
\concept{Work}
\concept{Work Orientation}
\keyword{variable}
