% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_504.R
\name{prideorg}
\alias{prideorg}
\title{I am proud working for my firm}
\source{
General Social Survey https://gss.norc.org
}
\description{
prideorg
}
\details{
Question To what extent do you agree or disagree with each of the following statements?
B. I am proud to be working for my firm or organization.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4586/vshow}{GSS Data Explorer page for \code{prideorg}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1967 \tab 192 \tab 417 \tab 165 \tab 32 \tab 5 \tab 9 \tab 45 \tab 2832 \cr
   2006 \tab 3492 \tab 369 \tab 473 \tab 120 \tab 37 \tab 16 \tab 1 \tab 2 \tab 4510 \cr
   2016 \tab 1927 \tab 339 \tab 428 \tab 126 \tab 29 \tab 9 \tab 2 \tab 7 \tab 2867 \cr
   Total \tab 7386 \tab 900 \tab 1318 \tab 411 \tab 98 \tab 30 \tab 12 \tab 54 \tab 10209 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{absent}},
\code{\link{chngwork}},
\code{\link{edcskill}},
\code{\link{helporg1}},
\code{\link{jobskill}},
\code{\link{proudwrk}},
\code{\link{stayorg3}},
\code{\link{useskill}},
\code{\link{wktenure}}
}
\concept{Work Orientation}
\keyword{variable}
