% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_86.R
\name{privacy}
\alias{privacy}
\title{Is r concerned about privacy?}
\source{
General Social Survey https://gss.norc.org
}
\description{
privacy
}
\details{
Question How concerned are you about threats to your personal privacy in America today? Would you say you are very concerned, somewhat concerned, only a little concerned, or not concerned at all?
}
\section{Values}{

\itemize{
\item \code{1} very concerned
\item \code{2} somewhat
\item \code{3} only a little
\item \code{4} not concerned
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/772/vshow}{GSS Data Explorer page for \code{privacy}}.
}

\seealso{
Other Core: 
\code{\link{bored}},
\code{\link{civic}},
\code{\link{feserve}},
\code{\link{imppol}},
\code{\link{meserve}},
\code{\link{pollgood}},
\code{\link{polltrue}},
\code{\link{rushed}},
\code{\link{stresswk}}
}
\concept{Core}
\keyword{variable}
