% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_498.R
\name{privgovt}
\alias{privgovt}
\title{R preference re: private public sector}
\source{
General Social Survey https://gss.norc.org
}
\description{
privgovt
}
\details{
Question Suppose you were working and could choose between different kinds of jobs. Which of the following would you personally choose?  I would advise...
Please CIRCLE ONE CODE FOR EACH.
D. Working in a private business (or) working for the government or civil service
}
\section{Values}{

\itemize{
\item \code{1} private firm
\item \code{2} govt
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4527/vshow}{GSS Data Explorer page for \code{privgovt}}.
}

\seealso{
Other Work Orientation: 
\code{\link{empself}},
\code{\link{findjob}},
\code{\link{indusoth}},
\code{\link{iwrkhard}},
\code{\link{smallbig}},
\code{\link{unbored}},
\code{\link{undk}},
\code{\link{wantjob}},
\code{\link{workweek}}
}
\concept{Work Orientation}
\keyword{variable}
