% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_644.R
\name{prntbhav}
\alias{prntbhav}
\title{Mother vs father teach child to behave}
\source{
General Social Survey https://gss.norc.org
}
\description{
prntbhav
}
\details{
Question (In your opinion, which parent should do the following? Who should...) ...teach the child how to behave?
}
\section{Values}{

\itemize{
\item \code{1} the mother mostly
\item \code{2} the mother somewhat more than the father
\item \code{3} the mother and father equally
\item \code{4} the father somewhat more than the mother
\item \code{5} the father mostly
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7771/vshow}{GSS Data Explorer page for \code{prntbhav}}.
}

\keyword{variable}
