% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_644.R
\name{prntmdl}
\alias{prntmdl}
\title{Mother vs father role model}
\source{
General Social Survey https://gss.norc.org
}
\description{
prntmdl
}
\details{
Question (In your opinion, which parent should do the following? Who should...) ...try to be a role model for the child?
}
\section{Values}{

\itemize{
\item \code{1} the mother mostly
\item \code{2} the mother somewhat more than the father
\item \code{3} the mother and father equally
\item \code{4} the father somewhat more than the mother
\item \code{5} the father mostly
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7773/vshow}{GSS Data Explorer page for \code{prntmdl}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab the mother mostly \tab the mother somewhat more than the father \tab the mother and father equally \tab the father somewhat more than the mother \tab the father mostly \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2383 \tab 8 \tab 2 \tab 1072 \tab 16 \tab 9 \tab 41 \tab 4 \tab 9 \tab 3544 \cr
   Total \tab 2383 \tab 8 \tab 2 \tab 1072 \tab 16 \tab 9 \tab 41 \tab 4 \tab 9 \tab 3544 \cr
}
}

\keyword{variable}
