% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_561.R
\name{prob3}
\alias{prob3}
\title{Third problem mentioned}
\source{
General Social Survey https://gss.norc.org
}
\description{
prob3
}
\details{
Question What are the most important problems that you and members of your household have had during the last 12 months?
RECORD VERBATIM.
PROBE FOR ADDITIONAL MENTIONS AND FOR SPECIFIC DETAILS. TRY NOT TO ACCEPT GENERAL RESPONSES LIKE "WORK" OR "CRIME."  PROBE TO FIND WHAT ABOUT "WORK" IS THE PROBLEM. FOR EXAMPLE, IS IT JOB INSECURITY, LOW PAY, ARGUMENTS WITH THE BOSS, OR WHAT? OR ABOUT "CRIME", WAS THERE A VICTIMIZATION?  WHO WAS VICTIMIZED AND IN WHAT WAY?
PROBE FOR ADDITIONAL PROBLEMS.
C. Third problem mentioned.
}
\section{Values}{

\itemize{
\item \verb{[995]} cant understand answer
\item \verb{[996]} no problems
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 1991, 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5134/vshow}{GSS Data Explorer page for \code{prob3}}.

Counts by year:\tabular{llllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 101 \tab 103 \tab 150 \tab 175 \tab 2 \tab 20 \tab 202 \tab 21 \tab 210 \tab 211 \tab 22 \tab 240 \tab 26 \tab 30 \tab 43 \tab 55 \tab 72 \tab 73 \tab 84 \tab 10 \tab 120 \tab 132 \tab 165 \tab 27 \tab 3 \tab 47 \tab Total \cr
   1991 \tab 1481 \tab 7 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   2004 \tab 2795 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2812 \cr
   Total \tab 4276 \tab 8 \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 2 \tab 3 \tab 4 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 4329 \cr
}
}

\seealso{
Other Negative Life Events: 
\code{\link{hlth1}},
\code{\link{hlth2}},
\code{\link{hlth3}},
\code{\link{hlth4}},
\code{\link{hlth5}},
\code{\link{hlth6}},
\code{\link{hlth7}},
\code{\link{prob2}},
\code{\link{prob4}}
}
\concept{Negative Life Events}
\keyword{variable}
