% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_561.R
\name{prob4}
\alias{prob4}
\title{Fourth problem mentioned}
\source{
General Social Survey https://gss.norc.org
}
\description{
prob4
}
\details{
Question What are the most important problems that you and members of your household have had during the last 12 months?
RECORD VERBATIM.
PROBE FOR ADDITIONAL MENTIONS AND FOR SPECIFIC DETAILS. TRY NOT TO ACCEPT GENERAL RESPONSES LIKE "WORK" OR "CRIME."  PROBE TO FIND WHAT ABOUT "WORK" IS THE PROBLEM. FOR EXAMPLE, IS IT JOB INSECURITY, LOW PAY, ARGUMENTS WITH THE BOSS, OR WHAT? OR ABOUT "CRIME", WAS THERE A VICTIMIZATION?  WHO WAS VICTIMIZED AND IN WHAT WAY?
PROBE FOR ADDITIONAL PROBLEMS.
D. Fourth problem mentioned.
}
\section{Values}{

\itemize{
\item \code{995} cant understand answer
\item \code{996} no problems
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab C \tab partial \cr
   2004 \tab C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5135/vshow}{GSS Data Explorer page for \code{prob4}}.
}

\seealso{
Other Negative Life Events: 
\code{\link{hlth1}},
\code{\link{hlth2}},
\code{\link{hlth3}},
\code{\link{hlth4}},
\code{\link{hlth5}},
\code{\link{hlth6}},
\code{\link{hlth7}},
\code{\link{prob2}},
\code{\link{prob3}}

Other Life Events: 
\code{\link{prob2}},
\code{\link{prob3}}

Other Trauma: 
\code{\link{hlth1}},
\code{\link{hlth2}},
\code{\link{hlth3}},
\code{\link{hlth4}},
\code{\link{hlth5}},
\code{\link{hlth6}},
\code{\link{hlth7}},
\code{\link{prob2}},
\code{\link{prob3}}
}
\concept{Life Events}
\concept{Negative Life Events}
\concept{Trauma}
\keyword{variable}
