% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_210.R
\name{probhlp1}
\alias{probhlp1}
\title{First mention to help problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
probhlp1
}
\details{
Question Problems often come up in life. Sometimes they're personal problems -- people are very unhappy, or nervous and irritable all the time. Sometimes there are problems in a marriage -- a husband and wife just can't get along with each other. Or, sometimes it's a personal problem with a child or a job. I'd like to ask you a few questions now about what you think a person might do to handle problems like this. For instance, let's suppose you had a lot of personal problems and you're very unhappy all the time. Let's suppose you've been that way for a long time, and it isn't getting any better. What do you think you'd Do about it?
\enumerate{
\item The sources of help R would seek -- First mention.
}
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1983/vshow}{GSS Data Explorer page for \code{probhlp1}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 100 \tab 103 \tab 104 \tab 105 \tab 106 \tab 107 \tab 108 \tab 190 \tab 200 \tab 201 \tab 202 \tab 290 \tab 300 \tab 390 \tab 400 \tab 401 \tab 402 \tab 403 \tab 490 \tab 499 \tab 500 \tab 502 \tab 503 \tab 505 \tab 506 \tab 509 \tab 590 \tab 602 \tab 608 \tab 690 \tab 700 \tab 800 \tab 980 \tab no answer \tab 102 \tab 910 \tab Total \cr
   1996 \tab 2168 \tab 16 \tab 8 \tab 1 \tab 2 \tab 27 \tab 9 \tab 2 \tab 5 \tab 51 \tab 10 \tab 1 \tab 2 \tab 1 \tab 2 \tab 41 \tab 3 \tab 2 \tab 27 \tab 14 \tab 27 \tab 25 \tab 8 \tab 1 \tab 3 \tab 158 \tab 1 \tab 1 \tab 1 \tab 1 \tab 3 \tab 2 \tab 6 \tab 1 \tab 274 \tab - \tab - \tab 2904 \cr
   2006 \tab 3776 \tab 11 \tab 3 \tab - \tab 1 \tab 27 \tab - \tab - \tab - \tab 23 \tab 1 \tab - \tab - \tab - \tab 1 \tab 19 \tab 2 \tab - \tab 96 \tab 40 \tab - \tab 15 \tab 7 \tab - \tab - \tab 111 \tab 11 \tab - \tab - \tab - \tab 2 \tab 10 \tab - \tab 88 \tab 40 \tab 1 \tab 225 \tab 4510 \cr
   Total \tab 5944 \tab 27 \tab 11 \tab 1 \tab 3 \tab 54 \tab 9 \tab 2 \tab 5 \tab 74 \tab 11 \tab 1 \tab 2 \tab 1 \tab 3 \tab 60 \tab 5 \tab 2 \tab 123 \tab 54 \tab 27 \tab 40 \tab 15 \tab 1 \tab 3 \tab 269 \tab 12 \tab 1 \tab 1 \tab 1 \tab 5 \tab 12 \tab 6 \tab 89 \tab 314 \tab 1 \tab 225 \tab 7414 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{charactr}},
\code{\link{imbalnce}},
\code{\link{probfix3}},
\code{\link{probhlp2}},
\code{\link{probhlp3}},
\code{\link{seriousp}},
\code{\link{stresses}},
\code{\link{vigversn}},
\code{\link{wayraise}}
}
\concept{Mental Health}
\keyword{variable}
