% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_210.R
\name{probhlp1}
\alias{probhlp1}
\title{First mention to help problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
probhlp1
}
\details{
Question Problems often come up in life. Sometimes they're personal problems -- people are very unhappy, or nervous and irritable all the time. Sometimes there are problems in a marriage -- a husband and wife just can't get along with each other. Or, sometimes it's a personal problem with a child or a job. I'd like to ask you a few questions now about what you think a person might do to handle problems like this. For instance, let's suppose you had a lot of personal problems and you're very unhappy all the time. Let's suppose you've been that way for a long time, and it isn't getting any better. What do you think you'd Do about it?
\enumerate{
\item The sources of help R would seek -- First mention.
}
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1983/vshow}{GSS Data Explorer page for \code{probhlp1}}.
}

\seealso{
Other Mental Health: 
\code{\link{charactr}},
\code{\link{imbalnce}},
\code{\link{probfix3}},
\code{\link{probhlp2}},
\code{\link{probhlp3}},
\code{\link{seriousp}},
\code{\link{stresses}},
\code{\link{vigversn}},
\code{\link{wayraise}}
}
\concept{Mental Health}
\keyword{variable}
