% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_135.R
\name{profits1}
\alias{profits1}
\title{Whether management only care about profit}
\source{
General Social Survey https://gss.norc.org
}
\description{
profits1
}
\details{
Question To what extent do you agree or disagree with the following statements?
Please tell me whether you strongly agree, agree, disagree, or strongly disagree:
A. The way most companies work, the only thing management cares about is profits, regardless of what workers want or need. Do you…
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1263/vshow}{GSS Data Explorer page for \code{profits1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1991 \tab - \tab 356 \tab 687 \tab 376 \tab 46 \tab 46 \tab 6 \tab 1517 \cr
   Total \tab 0 \tab 356 \tab 687 \tab 376 \tab 46 \tab 46 \tab 6 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{genecanx}},
\code{\link{genecany}},
\code{\link{genegets}},
\code{\link{merged}},
\code{\link{outofbiz}},
\code{\link{profits2}},
\code{\link{reorg}},
\code{\link{unpower}},
\code{\link{unprog}}
}
\concept{Work Organizations}
\keyword{variable}
