% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_135.R
\name{profits2}
\alias{profits2}
\title{Should co. pay workers more than shareholders}
\source{
General Social Survey https://gss.norc.org
}
\description{
profits2
}
\details{
Question To what extent do you agree or disagree with the following statements?
Please tell me whether you strongly agree, agree, disagree, or strongly disagree:
B. Corporations should pay more of their profits to workers and less to shareholders…
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1264/vshow}{GSS Data Explorer page for \code{profits2}}.
}

\seealso{
Other Work Organizations: 
\code{\link{genecanx}},
\code{\link{genecany}},
\code{\link{genegets}},
\code{\link{merged}},
\code{\link{outofbiz}},
\code{\link{profits1}},
\code{\link{reorg}},
\code{\link{unpower}},
\code{\link{unprog}}

Other Corporations: 
\code{\link{merged}},
\code{\link{outofbiz}},
\code{\link{profits1}},
\code{\link{reorg}}

Other Social Class: 
\code{\link{profits1}}
}
\concept{Corporations}
\concept{Social Class}
\concept{Work Organizations}
\keyword{variable}
