% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_436.R
\name{progtax}
\alias{progtax}
\title{How much income tax should rich pay?}
\source{
General Social Survey https://gss.norc.org
}
\description{
progtax
}
\details{
Question Some people think those with high incomes should pay a larger proportion (percentage) of their earnings in taxes than those who earn low incomes. Other people think that those with high incomes and those with low incomes should pay the same proportion (percentage) of their earnings in taxes. Do you think those with high incomes should pay a much larger proportion, pay a larger proportion, pay the same proportion as those who earn low incomes, pay a smaller proportion. pay a much smaller proportion?
}
\section{Values}{

\itemize{
\item \code{1} pay much more
\item \code{2} pay more
\item \code{3} all pay same
\item \code{4} pay less
\item \code{5} pay much less
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4004/vshow}{GSS Data Explorer page for \code{progtax}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{databank}},
\code{\link{eqincome}},
\code{\link{manhold}},
\code{\link{manread}},
\code{\link{mantail}},
\code{\link{mantap}},
\code{\link{opprof}},
\code{\link{oprich}},
\code{\link{verdict}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
