% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_139.R
\name{promoted}
\alias{promoted}
\title{Ever promoted by current employer}
\source{
General Social Survey https://gss.norc.org
}
\description{
promoted
}
\details{
Question Have you received any promotions while working for your present employer (IF SELF-EMPLOYED:  current organization)?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1300/vshow}{GSS Data Explorer page for \code{promoted}}.
}

\seealso{
Other Work Organizations: 
\code{\link{advances}},
\code{\link{futpromo}},
\code{\link{imppromo}},
\code{\link{jobcntrl}},
\code{\link{lottosay}},
\code{\link{numpromo}},
\code{\link{racpromo}},
\code{\link{sexpromo}},
\code{\link{wrkindep}}

Other Employment: 
\code{\link{advances}},
\code{\link{futpromo}},
\code{\link{imppromo}},
\code{\link{numpromo}}
}
\concept{Employment}
\concept{Work Organizations}
\keyword{variable}
