% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_308.R
\name{promtefr}
\alias{promtefr}
\title{Promotions are handled fairly}
\source{
General Social Survey https://gss.norc.org
}
\description{
promtefr
}
\details{
Question Now I''m going to read you another list of statements about your main job. For each, please tell me if the statement is very true, somewhat true, not too true, or not at all true with respect to the work you do.
I. Promotions are handled fairly
}
\section{Values}{

\itemize{
\item \verb{[1]} very true
\item \verb{[2]} somewhat true
\item \verb{[3]} not too true
\item \verb{[4]} not at all true
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2804/vshow}{GSS Data Explorer page for \code{promtefr}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very true \tab somewhat true \tab not too true \tab not at all true \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 507 \tab 622 \tab 275 \tab 254 \tab 108 \tab 30 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 469 \tab 629 \tab 247 \tab 207 \tab 153 \tab 29 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 329 \tab 443 \tab 161 \tab 149 \tab 63 \tab 42 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 368 \tab 474 \tab 191 \tab 123 \tab 79 \tab 15 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 444 \tab 566 \tab 178 \tab 129 \tab 88 \tab 43 \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 574 \tab 763 \tab 296 \tab 199 \tab 98 \tab 21 \tab 38 \tab 3544 \cr
   Total \tab 8345 \tab 2691 \tab 3497 \tab 1348 \tab 1061 \tab 589 \tab 180 \tab 38 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{condemnd}},
\code{\link{cowrkint}},
\code{\link{fringeok}},
\code{\link{haveinfo}},
\code{\link{hlpequip}},
\code{\link{jobsecok}},
\code{\link{opdevel}},
\code{\link{supcares}},
\code{\link{wkfreedm}}
}
\concept{Quality of Working Life}
\keyword{variable}
