% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_434.R
\name{protest4}
\alias{protest4}
\title{Allow occupation of govt office}
\source{
General Social Survey https://gss.norc.org
}
\description{
protest4
}
\details{
Question There are many ways people or organizations can protest against a government action they strongly oppose. Please show which you think should be allowed and which should not be allowed by circling a number after each question.
D. Occupying a government office and stopping work there for several days.
}
\section{Values}{

\itemize{
\item \code{1} definitely allowed
\item \code{2} probably allowed
\item \code{3} probably not allowed
\item \code{4} definitely not allowed
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3985/vshow}{GSS Data Explorer page for \code{protest4}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely allowed \tab probably allowed \tab probably not allowed \tab definitely not allowed \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 19 \tab 30 \tab 162 \tab 394 \tab 44 \tab 28 \tab 1534 \cr
   1990 \tab 155 \tab 43 \tab 98 \tab 326 \tab 643 \tab 81 \tab 26 \tab 1372 \cr
   Total \tab 1012 \tab 62 \tab 128 \tab 488 \tab 1037 \tab 125 \tab 54 \tab 2906 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{obeylaw}},
\code{\link{protest1}},
\code{\link{protest2}},
\code{\link{protest3}},
\code{\link{protest5}},
\code{\link{protest6}},
\code{\link{pubdef}},
\code{\link{pubecon}},
\code{\link{ratetone}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
