% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_531.R
\name{proudart}
\alias{proudart}
\title{How proud its achievements in the arts & lit.}
\source{
General Social Survey https://gss.norc.org
}
\description{
proudart
}
\details{
Question How proud are you of America in each of the following?
G. Its achievements in the arts and literature
}
\section{Values}{

\itemize{
\item \code{1} very proud
\item \code{2} somewhat proud
\item \code{3} not very proud
\item \code{4} not proud at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4846/vshow}{GSS Data Explorer page for \code{proudart}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very proud \tab somewhat proud \tab not very proud \tab not proud at all \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 386 \tab 714 \tab 112 \tab 32 \tab 91 \tab 32 \tab 2904 \cr
   2004 \tab 1596 \tab 469 \tab 581 \tab 72 \tab 12 \tab 81 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 421 \tab 622 \tab 83 \tab 18 \tab 126 \tab 4 \tab 2538 \cr
   Total \tab 4397 \tab 1276 \tab 1917 \tab 267 \tab 62 \tab 298 \tab 37 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amsports}},
\code{\link{ifwrong}},
\code{\link{prouddem}},
\code{\link{proudeco}},
\code{\link{proudmil}},
\code{\link{proudpol}},
\code{\link{proudsci}},
\code{\link{proudspt}},
\code{\link{proudsss}}
}
\concept{ISSP National Identity}
\keyword{variable}
