% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_531.R
\name{proudmil}
\alias{proudmil}
\title{How proud america's armed forces}
\source{
General Social Survey https://gss.norc.org
}
\description{
proudmil
}
\details{
Question How proud are you of America in each of the following?
H. America''s armed forces
}
\section{Values}{

\itemize{
\item \verb{[1]} very proud
\item \verb{[2]} somewhat proud
\item \verb{[3]} not very proud
\item \verb{[4]} not proud at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4847/vshow}{GSS Data Explorer page for \code{proudmil}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very proud \tab somewhat proud \tab not very proud \tab not proud at all \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 631 \tab 539 \tab 87 \tab 31 \tab 54 \tab 25 \tab 2904 \cr
   2004 \tab 1596 \tab 898 \tab 214 \tab 53 \tab 19 \tab 31 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 815 \tab 328 \tab 64 \tab 20 \tab 44 \tab 3 \tab 2538 \cr
   Total \tab 4397 \tab 2344 \tab 1081 \tab 204 \tab 70 \tab 129 \tab 29 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amsports}},
\code{\link{ifwrong}},
\code{\link{proudart}},
\code{\link{prouddem}},
\code{\link{proudeco}},
\code{\link{proudpol}},
\code{\link{proudsci}},
\code{\link{proudspt}},
\code{\link{proudsss}}
}
\concept{ISSP National Identity}
\keyword{variable}
