% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_531.R
\name{proudspt}
\alias{proudspt}
\title{How proud its achievements in sports}
\source{
General Social Survey https://gss.norc.org
}
\description{
proudspt
}
\details{
Question How proud are you of America in each of the following?
F. Its achievements in sports
}
\section{Values}{

\itemize{
\item \code{1} very proud
\item \code{2} somewhat proud
\item \code{3} not very proud
\item \code{4} not proud at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4845/vshow}{GSS Data Explorer page for \code{proudspt}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very proud \tab somewhat proud \tab not very proud \tab not proud at all \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 476 \tab 641 \tab 103 \tab 44 \tab 75 \tab 28 \tab 2904 \cr
   2004 \tab 1596 \tab 511 \tab 520 \tab 88 \tab 23 \tab 72 \tab 2 \tab 2812 \cr
   2014 \tab 1264 \tab 434 \tab 598 \tab 79 \tab 34 \tab 126 \tab 3 \tab 2538 \cr
   Total \tab 4397 \tab 1421 \tab 1759 \tab 270 \tab 101 \tab 273 \tab 33 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amsports}},
\code{\link{ifwrong}},
\code{\link{proudart}},
\code{\link{prouddem}},
\code{\link{proudeco}},
\code{\link{proudmil}},
\code{\link{proudpol}},
\code{\link{proudsci}},
\code{\link{proudsss}}
}
\concept{ISSP National Identity}
\keyword{variable}
