% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_504.R
\name{proudwrk}
\alias{proudwrk}
\title{I am proud of the type of work i do}
\source{
General Social Survey https://gss.norc.org
}
\description{
proudwrk
}
\details{
Question To what extent do you agree or disagree with each of the following statements?
E. I am proud of the type of work I do.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   B/C/- \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4589/vshow}{GSS Data Explorer page for \code{proudwrk}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1967 \tab 253 \tab 385 \tab 143 \tab 26 \tab 14 \tab 4 \tab 40 \tab 2832 \cr
   2016 \tab 1927 \tab 373 \tab 427 \tab 99 \tab 27 \tab 9 \tab - \tab 5 \tab 2867 \cr
   Total \tab 3894 \tab 626 \tab 812 \tab 242 \tab 53 \tab 23 \tab 4 \tab 45 \tab 5699 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{absent}},
\code{\link{chngwork}},
\code{\link{edcskill}},
\code{\link{helporg1}},
\code{\link{jobskill}},
\code{\link{prideorg}},
\code{\link{stayorg3}},
\code{\link{useskill}},
\code{\link{wktenure}}
}
\concept{Work Orientation}
\keyword{variable}
