% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_246.R
\name{prozac}
\alias{prozac}
\title{Have you ever heard of prozac}
\source{
General Social Survey https://gss.norc.org
}
\description{
prozac
}
\details{
Question Next, I''d like to ask you some questions about the medication called Prozac.
A. Have you ever heard of Prozac?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2220/vshow}{GSS Data Explorer page for \code{prozac}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 1191 \tab 161 \tab 29 \tab 6 \tab 2832 \cr
   2006 \tab 3073 \tab 1258 \tab 152 \tab 13 \tab 14 \tab 4510 \cr
   Total \tab 4518 \tab 2449 \tab 313 \tab 42 \tab 20 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{doc12}},
\code{\link{doc13}},
\code{\link{doc14}},
\code{\link{doc15}},
\code{\link{doc16}},
\code{\link{doc17}},
\code{\link{doc18}},
\code{\link{doc19}},
\code{\link{doc20}}
}
\concept{Mental Health}
\keyword{variable}
