% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_247.R
\name{prozfor2}
\alias{prozfor2}
\title{What can you tell me about prozac is used for}
\source{
General Social Survey https://gss.norc.org
}
\description{
prozfor2
}
\details{
Question What can you tell me about what Prozac is used for?
B. Second mention
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2224/vshow}{GSS Data Explorer page for \code{prozfor2}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 3 \tab 4 \tab 5 \tab 50 \tab 6 \tab 7 \tab 8 \tab 9 \tab no answer \tab 27 \tab 29 \tab 30 \tab 31 \tab 33 \tab 34 \tab don't know \tab Total \cr
   1998 \tab 2515 \tab 35 \tab 1 \tab 30 \tab 3 \tab 4 \tab 6 \tab 9 \tab 8 \tab 3 \tab 3 \tab 2 \tab 58 \tab 8 \tab 5 \tab 3 \tab 2 \tab 8 \tab 30 \tab 3 \tab 22 \tab 14 \tab 21 \tab 5 \tab 6 \tab 14 \tab 5 \tab 3 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2006 \tab 3252 \tab - \tab 20 \tab 4 \tab 25 \tab 18 \tab 11 \tab - \tab 7 \tab - \tab 10 \tab - \tab 10 \tab 13 \tab 16 \tab 3 \tab - \tab 7 \tab 2 \tab 5 \tab 3 \tab - \tab 1 \tab - \tab 78 \tab - \tab 1 \tab 1 \tab - \tab 2 \tab 8 \tab 5 \tab 2 \tab 2 \tab 5 \tab 999 \tab 4510 \cr
   Total \tab 5767 \tab 35 \tab 21 \tab 34 \tab 28 \tab 22 \tab 17 \tab 9 \tab 15 \tab 3 \tab 13 \tab 2 \tab 68 \tab 21 \tab 21 \tab 6 \tab 2 \tab 15 \tab 32 \tab 8 \tab 25 \tab 14 \tab 22 \tab 5 \tab 84 \tab 14 \tab 6 \tab 4 \tab 6 \tab 2 \tab 8 \tab 5 \tab 2 \tab 2 \tab 5 \tab 999 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{knwnproz}},
\code{\link{proz1}},
\code{\link{proz2}},
\code{\link{proz3}},
\code{\link{proz4}},
\code{\link{proz5}},
\code{\link{prozfor1}},
\code{\link{prozfor3}},
\code{\link{usedproz}}
}
\concept{Mental Health}
\keyword{variable}
