% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_247.R
\name{prozfor3}
\alias{prozfor3}
\title{What can you tell me about prozac is used for}
\source{
General Social Survey https://gss.norc.org
}
\description{
prozfor3
}
\details{
Question What can you tell me about what Prozac is used for?
C. Third Mention
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2225/vshow}{GSS Data Explorer page for \code{prozfor3}}.
}

\seealso{
Other Mental Health: 
\code{\link{knwnproz}},
\code{\link{proz1}},
\code{\link{proz2}},
\code{\link{proz3}},
\code{\link{proz4}},
\code{\link{proz5}},
\code{\link{prozfor1}},
\code{\link{prozfor2}},
\code{\link{usedproz}}
}
\concept{Mental Health}
\keyword{variable}
