% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_248.R
\name{prozslf3}
\alias{prozslf3}
\title{Take prozac if feel depressed sleepless}
\source{
General Social Survey https://gss.norc.org
}
\description{
prozslf3
}
\details{
Question How likely would you be to take doctor prescribed Prozac in the following situations?
C. ...because you were feeling depressed, tired, were having trouble sleeping and concentrating, and felt worthless.
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} mixed
\item \code{4} somewht unlikely
\item \code{5} very unlikely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2237/vshow}{GSS Data Explorer page for \code{prozslf3}}.
}

\seealso{
Other Mental Health: 
\code{\link{proz6}},
\code{\link{proz7}},
\code{\link{proz8}},
\code{\link{proz9}},
\code{\link{prozkid1}},
\code{\link{prozkid2}},
\code{\link{prozslf1}},
\code{\link{prozslf2}},
\code{\link{prozslf4}}

Other Mental Health: 
\code{\link{proz6}},
\code{\link{proz7}},
\code{\link{proz8}},
\code{\link{proz9}},
\code{\link{prozkid1}},
\code{\link{prozkid2}},
\code{\link{prozslf1}},
\code{\link{prozslf2}},
\code{\link{prozslf4}}
}
\concept{Mental Health}
\keyword{variable}
