% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_445.R
\name{prvdhlth}
\alias{prvdhlth}
\title{Who provides for sick people}
\source{
General Social Survey https://gss.norc.org
}
\description{
prvdhlth
}
\details{
Question (HAND CARD C3)
People have different opinions on who should provide services in America. Who do you think should primarily provide...
Healthcare for the sick?  Should it be...
}
\section{Values}{

\itemize{
\item \code{1} government
\item \code{2} private companies/for-profit organizations
\item \code{3} non-profit organizations/charities/cooperatives
\item \code{4} religious organizations
\item \code{5} family, relatives or friends
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab A/B \tab partial \cr
   2018 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6305/vshow}{GSS Data Explorer page for \code{prvdhlth}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gendereq}},
\code{\link{grnlaws}},
\code{\link{polint}},
\code{\link{prvdold}},
\code{\link{rprtst1}},
\code{\link{rprtst15}},
\code{\link{rprtst3}},
\code{\link{rprtst35}},
\code{\link{unionsok}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
