% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_445.R
\name{prvdold}
\alias{prvdold}
\title{Who provides for old people}
\source{
General Social Survey https://gss.norc.org
}
\description{
prvdold
}
\details{
Question (HAND CARD C3)
(People have different opinions on who should provide services in America. Who do you think should primarily provide...)
Care for old people?  (Should it be?)
}
\section{Values}{

\itemize{
\item \verb{[1]} government
\item \verb{[2]} private companies/for-profit organizations
\item \verb{[3]} non-profit organizations/charities/cooperatives
\item \verb{[4]} religious organizations
\item \verb{[5]} family, relatives or friends
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2016, 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6306/vshow}{GSS Data Explorer page for \code{prvdold}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab government \tab private companies/for-profit organizations \tab non-profit organizations/charities/cooperatives \tab religious organizations \tab family, relatives or friends \tab can't choose \tab no answer \tab Total \cr
   2016 \tab 1477 \tab 685 \tab 147 \tab 150 \tab 7 \tab 359 \tab 24 \tab 18 \tab 2867 \cr
   2018 \tab 1175 \tab 644 \tab 97 \tab 90 \tab 8 \tab 284 \tab 47 \tab 3 \tab 2348 \cr
   Total \tab 2652 \tab 1329 \tab 244 \tab 240 \tab 15 \tab 643 \tab 71 \tab 21 \tab 5215 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gendereq}},
\code{\link{grnlaws}},
\code{\link{polint}},
\code{\link{prvdhlth}},
\code{\link{rprtst1}},
\code{\link{rprtst15}},
\code{\link{rprtst3}},
\code{\link{rprtst35}},
\code{\link{unionsok}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
