% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_446.R
\name{prvdschl}
\alias{prvdschl}
\title{Who provides school educ}
\source{
General Social Survey https://gss.norc.org
}
\description{
prvdschl
}
\details{
Question (HAND CARD C3) 
(People have different opinions on who should provide services in America. Who do you think should primarily provide...)
School education for children?  (Should it be?)
}
\section{Values}{

\itemize{
\item \code{1} government
\item \code{2} private companies/for-profit organizations
\item \code{3} non-profit organizations/charities/cooperatives
\item \code{4} religious organizations
\item \code{5} family, relatives or friends
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6307/vshow}{GSS Data Explorer page for \code{prvdschl}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gvinflu1}},
\code{\link{gvinflu2}},
\code{\link{poleff11}},
\code{\link{poleff12}},
\code{\link{poleff13}},
\code{\link{poleff14}},
\code{\link{poleff15}},
\code{\link{poleff16}},
\code{\link{polgbeco}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
