% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_241.R
\name{psycmed3}
\alias{psycmed3}
\title{Taking these medicine interferes w daily act}
\source{
General Social Survey https://gss.norc.org
}
\description{
psycmed3
}
\details{
Question Please tell me how much you agree or disagree with the following statements about medicines prescribed by doctors to help people who are having problems with their emotions, nerves or their mental health:
C. Taking these medications interferes with daily activities.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2170/vshow}{GSS Data Explorer page for \code{psycmed3}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 99 \tab 398 \tab 310 \tab 335 \tab 62 \tab 179 \tab 4 \tab 2832 \cr
   2006 \tab 3073 \tab 110 \tab 508 \tab 344 \tab 324 \tab 58 \tab 79 \tab 14 \tab 4510 \cr
   Total \tab 4518 \tab 209 \tab 906 \tab 654 \tab 659 \tab 120 \tab 258 \tab 18 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{psycmed1}},
\code{\link{psycmed2}},
\code{\link{seenmntl}}
}
\concept{Mental Health}
\keyword{variable}
