% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_241.R
\name{psycmed3}
\alias{psycmed3}
\title{Taking these medicine interferes w daily act}
\source{
General Social Survey https://gss.norc.org
}
\description{
psycmed3
}
\details{
Question Please tell me how much you agree or disagree with the following statements about medicines prescribed by doctors to help people who are having problems with their emotions, nerves or their mental health:
C. Taking these medications interferes with daily activities.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2170/vshow}{GSS Data Explorer page for \code{psycmed3}}.
}

\seealso{
Other Mental Health: 
\code{\link{psycmed1}},
\code{\link{psycmed2}},
\code{\link{seenmntl}}
}
\concept{Mental Health}
\keyword{variable}
