% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_242.R
\name{psycmed4}
\alias{psycmed4}
\title{Helps people deal w day to day stresses}
\source{
General Social Survey https://gss.norc.org
}
\description{
psycmed4
}
\details{
Question Please tell me how much you agree or disagree with the following statements about medicines prescribed by doctors to help people who are having problems with their emotions, nerves or their mental health:
D. Taking these medications helps people deal with day-to-day stresses.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2171/vshow}{GSS Data Explorer page for \code{psycmed4}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 159 \tab 847 \tab 192 \tab 66 \tab 15 \tab 103 \tab 5 \tab 2832 \cr
   2006 \tab 3073 \tab 229 \tab 912 \tab 152 \tab 60 \tab 13 \tab 57 \tab 14 \tab 4510 \cr
   Total \tab 4518 \tab 388 \tab 1759 \tab 344 \tab 126 \tab 28 \tab 160 \tab 19 \tab 7342 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{psyckid1}},
\code{\link{psyckid2}},
\code{\link{psycmed5}},
\code{\link{psycmed6}},
\code{\link{psycmed7}},
\code{\link{usepsyc1}},
\code{\link{usepsyc2}},
\code{\link{usepsyc3}},
\code{\link{usepsyc4}}
}
\concept{Mental Health}
\keyword{variable}
