% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_242.R
\name{psycmed5}
\alias{psycmed5}
\title{Makes things easier in relations}
\source{
General Social Survey https://gss.norc.org
}
\description{
psycmed5
}
\details{
Question Please tell me how much you agree or disagree with the following statements about medicines prescribed by doctors to help people who are having problems with their emotions, nerves or their mental health:
E. Taking these medications makes things easier in relations with family and friends.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2172/vshow}{GSS Data Explorer page for \code{psycmed5}}.
}

\seealso{
Other Mental Health: 
\code{\link{psyckid1}},
\code{\link{psyckid2}},
\code{\link{psycmed4}},
\code{\link{psycmed6}},
\code{\link{psycmed7}},
\code{\link{usepsyc1}},
\code{\link{usepsyc2}},
\code{\link{usepsyc3}},
\code{\link{usepsyc4}}

Other Mental Health: 
\code{\link{psyckid1}},
\code{\link{psyckid2}},
\code{\link{psycmed4}},
\code{\link{psycmed6}},
\code{\link{psycmed7}},
\code{\link{usepsyc1}},
\code{\link{usepsyc2}},
\code{\link{usepsyc3}},
\code{\link{usepsyc4}}
}
\concept{Mental Health}
\keyword{variable}
