% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_635.R
\name{ptnrornt}
\alias{ptnrornt}
\title{Recent partner's sexual orientation}
\source{
General Social Survey https://gss.norc.org
}
\description{
ptnrornt
}
\details{
Question To the best of your knowledge, which of the following best describes how your most recent sexual partner thinks of themselves?
}
\section{Values}{

\itemize{
\item \code{1} gay, lesbian, or homosexual
\item \code{2} bisexual
\item \code{3} heterosexual or straight
\item \code{4} i have never had a sexual partner
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab B/C \tab partial \cr
   2022 \tab B/C \tab partial \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7559/vshow}{GSS Data Explorer page for \code{ptnrornt}}.
}

\keyword{variable}
