% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_635.R
\name{ptnrornt}
\alias{ptnrornt}
\title{Recent partner's sexual orientation}
\source{
General Social Survey https://gss.norc.org
}
\description{
ptnrornt
}
\details{
Question To the best of your knowledge, which of the following best describes how your most recent sexual partner thinks of themselves?
}
\section{Values}{

\itemize{
\item \verb{[1]} gay, lesbian, or homosexual
\item \verb{[2]} bisexual
\item \verb{[3]} heterosexual or straight
\item \verb{[4]} i have never had a sexual partner
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7559/vshow}{GSS Data Explorer page for \code{ptnrornt}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab gay, lesbian, or homosexual \tab bisexual \tab heterosexual or straight \tab i have never had a sexual partner \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 1686 \tab 74 \tab 67 \tab 2020 \tab 60 \tab 68 \tab 7 \tab 50 \tab 4032 \cr
   2022 \tab 1900 \tab 47 \tab 64 \tab 1330 \tab 54 \tab 86 \tab 12 \tab 51 \tab 3544 \cr
   Total \tab 3586 \tab 121 \tab 131 \tab 3350 \tab 114 \tab 154 \tab 19 \tab 101 \tab 7576 \cr
}
}

\keyword{variable}
