% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_635.R
\name{ptnrsxbrth}
\alias{ptnrsxbrth}
\title{Recent partner's sex at birth}
\source{
General Social Survey https://gss.norc.org
}
\description{
ptnrsxbrth
}
\details{
Question To the best of your knowledge, was your most recent sexual partner's sex recorded as male or female at birth?
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7560/vshow}{GSS Data Explorer page for \code{ptnrsxbrth}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab male \tab female \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 1746 \tab 1228 \tab 974 \tab 40 \tab 5 \tab 39 \tab 4032 \cr
   2022 \tab 1954 \tab 832 \tab 670 \tab 46 \tab 9 \tab 33 \tab 3544 \cr
   Total \tab 3700 \tab 2060 \tab 1644 \tab 86 \tab 14 \tab 72 \tab 7576 \cr
}
}

\keyword{variable}
