% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_635.R
\name{ptnrsxnow}
\alias{ptnrsxnow}
\title{Recent partner's sex now}
\source{
General Social Survey https://gss.norc.org
}
\description{
ptnrsxnow
}
\details{
Question To the best of your knowledge, does your most recent sexual partner describe themselves as male, female, or transgender?
}
\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{3} transgender
\item \code{4} none of these
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab B/C \tab partial \cr
   2022 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7561/vshow}{GSS Data Explorer page for \code{ptnrsxnow}}.
}

\keyword{variable}
