% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_635.R
\name{ptnrsxnow}
\alias{ptnrsxnow}
\title{Recent partner's sex now}
\source{
General Social Survey https://gss.norc.org
}
\description{
ptnrsxnow
}
\details{
Question To the best of your knowledge, does your most recent sexual partner describe themselves as male, female, or transgender?
}
\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{3} transgender
\item \code{4} none of these
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7561/vshow}{GSS Data Explorer page for \code{ptnrsxnow}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab male \tab female \tab transgender \tab none of these \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 1747 \tab 1227 \tab 954 \tab 5 \tab 12 \tab 43 \tab 4 \tab 40 \tab 4032 \cr
   2022 \tab 1955 \tab 821 \tab 668 \tab 4 \tab 4 \tab 43 \tab 9 \tab 40 \tab 3544 \cr
   Total \tab 3702 \tab 2048 \tab 1622 \tab 9 \tab 16 \tab 86 \tab 13 \tab 80 \tab 7576 \cr
}
}

\keyword{variable}
