% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_686.R
\name{ptnrsxnow1}
\alias{ptnrsxnow1}
\title{Recent partner's sex now (3 category)}
\source{
General Social Survey https://gss.norc.org
}
\description{
ptnrsxnow1
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{3} redacted/none of these
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
