% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_523.R
\name{pubdecid}
\alias{pubdecid}
\title{Govt should let peop decide how to protect envir}
\source{
General Social Survey https://gss.norc.org
}
\description{
pubdecid
}
\details{
Question A. If you had to choose, which one of the following would be closest to your views?  Government should let ordinary people decide for themselves how to protect the environment, even if it means they don't always dothe right thing, or government should pass laws to make ordinary people protect the environment, even if it interferes with people's right to make their own decisions.
}
\section{Values}{

\itemize{
\item \code{1} gov. let people decide
\item \code{2} gov. make laws
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2010 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4766/vshow}{GSS Data Explorer page for \code{pubdecid}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{busdecid}},
\code{\link{chemfam}},
\code{\link{chemgen}},
\code{\link{indusfam}},
\code{\link{indusgen}},
\code{\link{tempfam}},
\code{\link{tempgen}},
\code{\link{waterfam}},
\code{\link{watergen}}
}
\concept{ISSP Environment}
\keyword{variable}
