% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_434.R
\name{pubdef}
\alias{pubdef}
\title{Should media publish defense plans}
\source{
General Social Survey https://gss.norc.org
}
\description{
pubdef
}
\details{
Question A. Suppose a newspaper got hold of confidential government papers about defense plans and wanted to publish them.
Should the newspaper be allowed to publish the papers? Or should the government have the power to prevent publication?
}
\section{Values}{

\itemize{
\item \code{1} allow publication
\item \code{2} stop publication
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3979/vshow}{GSS Data Explorer page for \code{pubdef}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{obeylaw}},
\code{\link{protest1}},
\code{\link{protest2}},
\code{\link{protest3}},
\code{\link{protest4}},
\code{\link{protest5}},
\code{\link{protest6}},
\code{\link{pubecon}},
\code{\link{ratetone}}

Other Authority: 
\code{\link{obeylaw}},
\code{\link{pubecon}}

Other Civil Liberties: 
\code{\link{pubecon}}

Other Media: 
\code{\link{pubecon}}
}
\concept{Authority}
\concept{Civil Liberties}
\concept{ISSP Role of Gov't}
\concept{Media}
\keyword{variable}
