% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_434.R
\name{pubecon}
\alias{pubecon}
\title{Should media publish economic plans?}
\source{
General Social Survey https://gss.norc.org
}
\description{
pubecon
}
\details{
Question B. Now suppose the confidential government papers were about economic plans.
Should the newspaper be allowed to publish the papers? Or should the government have the power to prevent publication?
}
\section{Values}{

\itemize{
\item \verb{[1]} allow publication
\item \verb{[2]} stop publication
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3980/vshow}{GSS Data Explorer page for \code{pubecon}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab allow publication \tab stop publication \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 359 \tab 227 \tab 70 \tab 21 \tab 1534 \cr
   Total \tab 857 \tab 359 \tab 227 \tab 70 \tab 21 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{obeylaw}},
\code{\link{protest1}},
\code{\link{protest2}},
\code{\link{protest3}},
\code{\link{protest4}},
\code{\link{protest5}},
\code{\link{protest6}},
\code{\link{pubdef}},
\code{\link{ratetone}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
