% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_571.R
\name{punenmy}
\alias{punenmy}
\title{Punishment selling info to foreign enemies}
\source{
General Social Survey https://gss.norc.org
}
\description{
punenmy
}
\details{
Question I’m going to ask you about several offenses. For each case, indicate what you would consider to be the appropriate punishment when the offender is a military service member or government employee.
C. Stealing and selling copies of secret codes and other classified intelligence materials to an agent of a hostile foreign government.
}
\section{Values}{

\itemize{
\item \code{1} reprimand
\item \code{2} dismissal from the military/fired by the government
\item \code{3} one year in prison
\item \code{4} 5-9 years in prison
\item \code{5} 10-20 years in prison
\item \code{6} life in prison
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5234/vshow}{GSS Data Explorer page for \code{punenmy}}.
}

\seealso{
Other National Security: 
\code{\link{comdata}},
\code{\link{comemail}},
\code{\link{comporn}},
\code{\link{comsnoop}},
\code{\link{comsteal}},
\code{\link{comsys}},
\code{\link{punfrnd}},
\code{\link{punleak}},
\code{\link{puntrck}}

Other United States: 
\code{\link{comdata}},
\code{\link{comemail}},
\code{\link{comporn}},
\code{\link{comsnoop}},
\code{\link{comsteal}},
\code{\link{comsys}},
\code{\link{punfrnd}},
\code{\link{punleak}},
\code{\link{puntrck}}
}
\concept{National Security}
\concept{United States}
\keyword{variable}
