% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_335.R
\name{putdown}
\alias{putdown}
\title{Ppl at work treat r in a manner putting r down}
\source{
General Social Survey https://gss.norc.org
}
\description{
putdown
}
\details{
Question D. People at work treat me in a manner that puts me down or address me in unprofessional terms, either publicly or privately.
}
\section{Values}{

\itemize{
\item \verb{[1]} often
\item \verb{[2]} sometimes
\item \verb{[3]} rarely
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3001/vshow}{GSS Data Explorer page for \code{putdown}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab often \tab sometimes \tab rarely \tab never \tab don't know \tab no answer \tab Total \cr
   2004 \tab 945 \tab 53 \tab 228 \tab 410 \tab 1114 \tab 21 \tab 41 \tab 2812 \cr
   2012 \tab 795 \tab 38 \tab 134 \tab 315 \tab 645 \tab 14 \tab 33 \tab 1974 \cr
   Total \tab 1740 \tab 91 \tab 362 \tab 725 \tab 1759 \tab 35 \tab 74 \tab 4786 \cr
}
}

\seealso{
Other Workplace Conflict: 
\code{\link{difstand}},
\code{\link{lackinfo}},
\code{\link{othcredt}},
\code{\link{perspace}},
\code{\link{rptprobs}}
}
\concept{Workplace Conflict}
\keyword{variable}
