% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_142.R
\name{quality}
\alias{quality}
\title{How easy your supervisor know how good you are}
\source{
General Social Survey https://gss.norc.org
}
\description{
quality
}
\details{
Question How easy or hard would you say it is for you supervisor to evaluate the quality of the work that is done by a person in a job like yours?  Would you say very easy, somewhat easy, somewhat hard, or very hard?
}
\section{Values}{

\itemize{
\item \code{1} very easy
\item \code{2} somewhat easy
\item \code{3} somewhat hard
\item \code{4} very hard
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1337/vshow}{GSS Data Explorer page for \code{quality}}.
}

\seealso{
Other Work Organizations: 
\code{\link{jobjudge}},
\code{\link{orgmoney}},
\code{\link{othduty5}},
\code{\link{othduty6}},
\code{\link{othduty7}},
\code{\link{quantity}},
\code{\link{totmoney}},
\code{\link{wrkwell}},
\code{\link{youmoney}}
}
\concept{Work Organizations}
\keyword{variable}
