% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_48.R
\name{raccare}
\alias{raccare}
\title{R's concern about racial issues}
\source{
General Social Survey https://gss.norc.org
}
\description{
raccare
}
\details{
Question How concerned are you personally about race relations?  Are you very concerned, somewhat concerned, not very concerned, or not concerned at all?
}
\section{Values}{

\itemize{
\item \code{1} very concerned
\item \code{2} somewhat concerned
\item \code{3} not very concerned
\item \code{4} not concerned at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1983 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/418/vshow}{GSS Data Explorer page for \code{raccare}}.
}

\seealso{
Other Core: 
\code{\link{affrmact}},
\code{\link{blksimp}},
\code{\link{racfirm}},
\code{\link{racgive}},
\code{\link{racinfo}},
\code{\link{racjoin}},
\code{\link{racthink}},
\code{\link{racwrite}},
\code{\link{wrkwayup}}

Other Split Ballots: 
\code{\link{racfirm}},
\code{\link{racgive}},
\code{\link{racinfo}},
\code{\link{racjoin}},
\code{\link{racthink}},
\code{\link{racwrite}}
}
\concept{Core}
\concept{Split Ballots}
\keyword{variable}
