% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_81.R
\name{racchng}
\alias{racchng}
\title{Try to integrate social club}
\source{
General Social Survey https://gss.norc.org
}
\description{
racchng
}
\details{
Question If you and your friends belonged to a social club that would not let (Whites/Negroes/Blacks) join, would you try to change the rules so that (Whites/(Negroes/Blacks/African-Americans) could join?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} wdnt belong
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1977, 1985, 1986 \cr
   A/B \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/726/vshow}{GSS Data Explorer page for \code{racchng}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab wdnt belong \tab don't know \tab no answer \tab Total \cr
   1977 \tab 176 \tab 527 \tab 723 \tab 8 \tab 91 \tab 5 \tab 1530 \cr
   1985 \tab - \tab 800 \tab 637 \tab - \tab 94 \tab 3 \tab 1534 \cr
   1986 \tab - \tab 906 \tab 474 \tab - \tab 79 \tab 11 \tab 1470 \cr
   1988 \tab 493 \tab 558 \tab 342 \tab - \tab 78 \tab 10 \tab 1481 \cr
   1989 \tab 531 \tab 590 \tab 330 \tab - \tab 72 \tab 14 \tab 1537 \cr
   1990 \tab 444 \tab 540 \tab 296 \tab - \tab 85 \tab 7 \tab 1372 \cr
   1991 \tab 493 \tab 638 \tab 287 \tab - \tab 93 \tab 6 \tab 1517 \cr
   1993 \tab 526 \tab 693 \tab 300 \tab - \tab 76 \tab 11 \tab 1606 \cr
   1994 \tab 2485 \tab 306 \tab 148 \tab - \tab 50 \tab 3 \tab 2992 \cr
   Total \tab 5148 \tab 5558 \tab 3537 \tab 8 \tab 718 \tab 70 \tab 15039 \cr
}
}

\seealso{
Other Core: 
\code{\link{racavoid}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racquit}},
\code{\link{racsups}},
\code{\link{racteach}}
}
\concept{Core}
\keyword{variable}
