% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_46.R
\name{racchurh}
\alias{racchurh}
\title{Attend church with other race}
\source{
General Social Survey https://gss.norc.org
}
\description{
racchurh
}
\details{
Question Do (Blacks/Negroes/African-Americans)/Whites attend the church that you, yourself, attend most often, or not?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} no church
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1978 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab B/C \tab partial \cr
   1989 \tab B/C \tab partial \cr
   1990 \tab B/C \tab partial \cr
   1991 \tab B/C \tab partial \cr
   1993 \tab B/C \tab partial \cr
   1994 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/402/vshow}{GSS Data Explorer page for \code{racchurh}}.
}

\seealso{
Other Core: 
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{color}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{racjob}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racpres}}

Other Race: 
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{color}},
\code{\link{rachaf}},
\code{\link{rachisch}},
\code{\link{racjob}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racpres}}
}
\concept{Core}
\concept{Race}
\keyword{variable}
