% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_663.R
\name{racdif2x}
\alias{racdif2x}
\title{Differences due to in-born learning ability}
\source{
General Social Survey https://gss.norc.org
}
\description{
racdif2x
}
\details{
Question
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7964/vshow}{GSS Data Explorer page for \code{racdif2x}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab skipped on web \tab Total \cr
   2022 \tab 2339 \tab 87 \tab 1093 \tab 24 \tab 1 \tab 3544 \cr
   Total \tab 2339 \tab 87 \tab 1093 \tab 24 \tab 1 \tab 3544 \cr
}
}

\keyword{variable}
