% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_683.R
\name{racdif5x}
\alias{racdif5x}
\title{Differences due mainly to upbringing (longer intro)}
\source{
General Social Survey https://gss.norc.org
}
\description{
racdif5x
}
\details{
Question On the average Black people have worse jobs, income, and housing than white people.Do you think these differences are . . .
E. Mainly due to upbringing?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8203/vshow}{GSS Data Explorer page for \code{racdif5x}}.
}

\seealso{
Other Core: 
\code{\link{family16sex}},
\code{\link{femsee1}},
\code{\link{femself1}},
\code{\link{mascsee1}},
\code{\link{mascself1}},
\code{\link{sexnow2}}
}
\concept{Core}
\keyword{variable}
