% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_13.R
\name{race}
\alias{race}
\title{Race of respondent}
\source{
General Social Survey https://gss.norc.org
}
\description{
race
}
\details{
Question CODE WITHOUT ASKING ONLY IF THERE IS NO DOUBT IN YOUR MIND.
What race do you consider yourself? RECORD VERBATIM AND CODE.
}
\section{Values}{

\itemize{
\item \code{1} white
\item \code{2} black
\item \code{3} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/82/vshow}{GSS Data Explorer page for \code{race}}.

Counts by year:\tabular{llllll}{
   year \tab white \tab black \tab other \tab iap \tab Total \cr
   1972 \tab 1348 \tab 261 \tab 4 \tab - \tab 1613 \cr
   1973 \tab 1308 \tab 183 \tab 13 \tab - \tab 1504 \cr
   1974 \tab 1304 \tab 173 \tab 7 \tab - \tab 1484 \cr
   1975 \tab 1323 \tab 163 \tab 4 \tab - \tab 1490 \cr
   1976 \tab 1361 \tab 129 \tab 9 \tab - \tab 1499 \cr
   1977 \tab 1339 \tab 176 \tab 15 \tab - \tab 1530 \cr
   1978 \tab 1358 \tab 158 \tab 16 \tab - \tab 1532 \cr
   1980 \tab 1318 \tab 140 \tab 10 \tab - \tab 1468 \cr
   1982 \tab 1323 \tab 510 \tab 27 \tab - \tab 1860 \cr
   1983 \tab 1416 \tab 165 \tab 18 \tab - \tab 1599 \cr
   1984 \tab 1251 \tab 170 \tab 52 \tab - \tab 1473 \cr
   1985 \tab 1338 \tab 152 \tab 44 \tab - \tab 1534 \cr
   1986 \tab 1249 \tab 184 \tab 37 \tab - \tab 1470 \cr
   1987 \tab 1222 \tab 544 \tab 53 \tab - \tab 1819 \cr
   1988 \tab 1234 \tab 186 \tab 61 \tab - \tab 1481 \cr
   1989 \tab 1319 \tab 157 \tab 61 \tab - \tab 1537 \cr
   1990 \tab 1150 \tab 159 \tab 63 \tab - \tab 1372 \cr
   1991 \tab 1264 \tab 204 \tab 49 \tab - \tab 1517 \cr
   1993 \tab 1347 \tab 179 \tab 80 \tab - \tab 1606 \cr
   1994 \tab 2483 \tab 388 \tab 121 \tab - \tab 2992 \cr
   1996 \tab 2349 \tab 402 \tab 153 \tab - \tab 2904 \cr
   1998 \tab 2241 \tab 400 \tab 191 \tab - \tab 2832 \cr
   2000 \tab 2213 \tab 429 \tab 175 \tab - \tab 2817 \cr
   2002 \tab 2188 \tab 410 \tab 167 \tab - \tab 2765 \cr
   2004 \tab 2234 \tab 377 \tab 201 \tab - \tab 2812 \cr
   2006 \tab 3284 \tab 634 \tab 592 \tab - \tab 4510 \cr
   2008 \tab 1559 \tab 281 \tab 183 \tab - \tab 2023 \cr
   2010 \tab 1550 \tab 311 \tab 183 \tab - \tab 2044 \cr
   2012 \tab 1477 \tab 301 \tab 196 \tab - \tab 1974 \cr
   2014 \tab 1890 \tab 386 \tab 262 \tab - \tab 2538 \cr
   2016 \tab 2100 \tab 490 \tab 277 \tab - \tab 2867 \cr
   2018 \tab 1693 \tab 385 \tab 270 \tab - \tab 2348 \cr
   2021 \tab 3110 \tab 463 \tab 405 \tab 54 \tab 4032 \cr
   2022 \tab 2514 \tab 565 \tab 412 \tab 53 \tab 3544 \cr
   Total \tab 57657 \tab 10215 \tab 4411 \tab 107 \tab 72390 \cr
}
}

\seealso{
Other Core: 
\code{\link{cobarate}},
\code{\link{coeftotlt}},
\code{\link{cogradtounder}},
\code{\link{mobile16}},
\code{\link{reg16}},
\code{\link{res16}},
\code{\link{sex}},
\code{\link{spbarate}},
\code{\link{spgradtounder}}
}
\concept{Core}
\keyword{variable}
