% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_626.R
\name{raceacs13}
\alias{raceacs13}
\title{Samoan}
\source{
General Social Survey https://gss.norc.org
}
\description{
raceacs13
}
\details{
Question What is your race? Indicate one or more races that you consider yourself to be.  Samoan
As of 2024, RACEACS13 has been recoded into RACEACS14 as "Other Pacific Islander."
}
\section{Values}{

\itemize{
\item \verb{[0]} no
\item \verb{[1]} yes
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7462/vshow}{GSS Data Explorer page for \code{raceacs13}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab no \tab yes \tab don't know \tab no answer \tab refused \tab Total \cr
   2021 \tab - \tab 4013 \tab 4 \tab 6 \tab 9 \tab - \tab 4032 \cr
   2022 \tab - \tab 3500 \tab 3 \tab 8 \tab 18 \tab 15 \tab 3544 \cr
   Total \tab 0 \tab 7513 \tab 7 \tab 14 \tab 27 \tab 15 \tab 7576 \cr
}
}

\seealso{
Other Core: 
\code{\link{raceacs12}},
\code{\link{raceacs14}},
\code{\link{raceacs15}}
}
\concept{Core}
\keyword{variable}
