% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_626.R
\name{raceacs16}
\alias{raceacs16}
\title{Hispanic}
\source{
General Social Survey https://gss.norc.org
}
\description{
raceacs16
}
\details{
Question What is your race? Indicate one or more races that you consider yourself to be.  Hispanic or Latino
}
\section{Values}{

\itemize{
\item \verb{[0]} no
\item \verb{[1]} yes
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7465/vshow}{GSS Data Explorer page for \code{raceacs16}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab no \tab yes \tab don't know \tab no answer \tab refused \tab Total \cr
   2021 \tab - \tab 3858 \tab 169 \tab 4 \tab 1 \tab - \tab 4032 \cr
   2022 \tab - \tab 3346 \tab 157 \tab 8 \tab 18 \tab 15 \tab 3544 \cr
   Total \tab 0 \tab 7204 \tab 326 \tab 12 \tab 19 \tab 15 \tab 7576 \cr
}
}

\keyword{variable}
