% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_575.R
\name{racecen3}
\alias{racecen3}
\title{What is r's race, third mention}
\source{
General Social Survey https://gss.norc.org
}
\description{
racecen3
}
\details{
Question 1602. What is your race? Indicate one or more races that you consider yourself to be.
}
\section{Overview}{

For further details see the \href{https://gss.norc.org/get-documentation}{official GSS documentation}.

Counts by year:\tabular{lllllllllllll}{
   year \tab not available in this year \tab asian \tab black or african american \tab iap \tab american indian or alaska native \tab native hawaiian or other pacific islander \tab no answer \tab some other race \tab white \tab don't know \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab 1 \tab 1 \tab 2815 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 6 \tab 1 \tab 2747 \tab 6 \tab 1 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 2 \tab 1 \tab 2802 \tab 5 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab 2812 \cr
   2006 \tab - \tab 1 \tab 2 \tab 4476 \tab 10 \tab 2 \tab - \tab 3 \tab 12 \tab 4 \tab - \tab 4510 \cr
   2008 \tab - \tab 1 \tab 2 \tab 2011 \tab 3 \tab - \tab - \tab 1 \tab 3 \tab 2 \tab - \tab 2023 \cr
   2010 \tab - \tab 3 \tab - \tab 2025 \tab 5 \tab 2 \tab - \tab 1 \tab 8 \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 1 \tab 1 \tab 1964 \tab 5 \tab 1 \tab - \tab - \tab 2 \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 4 \tab - \tab 2519 \tab 8 \tab 1 \tab - \tab 3 \tab 3 \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 8 \tab 2 \tab 2830 \tab 20 \tab - \tab - \tab 1 \tab 6 \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 3 \tab 2 \tab 2322 \tab 5 \tab 1 \tab - \tab 10 \tab 5 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 3 \tab 3 \tab 4011 \tab 5 \tab 1 \tab - \tab 1 \tab 8 \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab 1 \tab 1 \tab 3525 \tab 4 \tab 2 \tab - \tab 2 \tab 9 \tab - \tab - \tab 3544 \cr
   2024 \tab - \tab 3 \tab 1 \tab 3249 \tab 4 \tab 1 \tab 26 \tab - \tab 12 \tab 9 \tab 4 \tab 3309 \cr
   Total \tab 38116 \tab 37 \tab 17 \tab 37296 \tab 80 \tab 12 \tab 29 \tab 23 \tab 69 \tab 16 \tab 4 \tab 75699 \cr
}
}

\section{Values}{

\itemize{
\item \code{1} white
\item \code{2} black or african american
\item \code{3} american indian or alaska native
\item \code{4} asian
\item \code{5} native hawaiian or other pacific islander
\item \code{6} some other race
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
