% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_575.R
\name{racedbtf}
\alias{racedbtf}
\title{Certainty of observed race}
\source{
General Social Survey https://gss.norc.org
}
\description{
racedbtf
}
\details{
Question B. FOR THE RACIAL CLASSIFICATION YOU JUST ASSIGNED IN "A" INDICATE HOW SURE YOU WERE OF RESPONDENT'S RACE.
}
\section{Values}{

\itemize{
\item \code{1} no doubt
\item \code{2} some doubt, pretty sure
\item \code{3} a lot of doubt, pretty unsure
\item \code{4} completely unsure
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5274/vshow}{GSS Data Explorer page for \code{racedbtf}}.
}

\seealso{
Other Ethnicity: 
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{hispanic}},
\code{\link{racecen1}},
\code{\link{racecen2}},
\code{\link{racecen3}},
\code{\link{racesee}},
\code{\link{raceself}}

Other Race: 
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{hispanic}},
\code{\link{racecen1}},
\code{\link{racecen2}},
\code{\link{racecen3}},
\code{\link{racesee}},
\code{\link{raceself}}
}
\concept{Ethnicity}
\concept{Race}
\keyword{variable}
