% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_575.R
\name{racedbtf}
\alias{racedbtf}
\title{Certainty of observed race}
\source{
General Social Survey https://gss.norc.org
}
\description{
racedbtf
}
\details{
Question B. FOR THE RACIAL CLASSIFICATION YOU JUST ASSIGNED IN "A" INDICATE HOW SURE YOU WERE OF RESPONDENT'S RACE.
}
\section{Values}{

\itemize{
\item \verb{[1]} no doubt
\item \verb{[2]} some doubt, pretty sure
\item \verb{[3]} a lot of doubt, pretty unsure
\item \verb{[4]} completely unsure
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5274/vshow}{GSS Data Explorer page for \code{racedbtf}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab no doubt \tab some doubt, pretty sure \tab a lot of doubt, pretty unsure \tab completely unsure \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 1334 \tab 58 \tab 1 \tab 10 \tab 57 \tab 2904 \cr
   2000 \tab - \tab 2543 \tab 127 \tab 8 \tab 9 \tab 130 \tab 2817 \cr
   Total \tab 1444 \tab 3877 \tab 185 \tab 9 \tab 19 \tab 187 \tab 5721 \cr
}
}

\seealso{
Other Ethnicity: 
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{hispanic}},
\code{\link{racecen1}},
\code{\link{racecen2}},
\code{\link{racecen3}},
\code{\link{racesee}},
\code{\link{raceself}}
}
\concept{Ethnicity}
\keyword{variable}
