% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_642.R
\name{racerank2}
\alias{racerank2}
\title{Second race selected}
\source{
General Social Survey https://gss.norc.org
}
\description{
racerank2
}
\details{
Question After your primary racial identity, which of these races do you identify most with?
}
\section{Values}{

\itemize{
\item \code{1} white
\item \code{2} black or african-american
\item \code{3} american indian or alaska native
\item \code{4} asian indian
\item \code{5} chinese
\item \code{6} filipino
\item \code{7} japanese
\item \code{8} korean
\item \code{9} vietnamese
\item \code{10} other asian
\item \code{11} native hawaiian
\item \code{12} guamanian or chamorro
\item \code{13} samoan
\item \code{14} other pacific islander
\item \code{15} some other race
\item \code{16} hispanic
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7752/vshow}{GSS Data Explorer page for \code{racerank2}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab iap \tab white \tab black or african-american \tab american indian or alaska native \tab asian indian \tab chinese \tab filipino \tab japanese \tab vietnamese \tab other asian \tab native hawaiian \tab samoan \tab other pacific islander \tab some other race \tab hispanic \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 3332 \tab 52 \tab 16 \tab 72 \tab 3 \tab 7 \tab 3 \tab 3 \tab 2 \tab 3 \tab 3 \tab 1 \tab 1 \tab 7 \tab 8 \tab 9 \tab 20 \tab 2 \tab 3544 \cr
   Total \tab 3332 \tab 52 \tab 16 \tab 72 \tab 3 \tab 7 \tab 3 \tab 3 \tab 2 \tab 3 \tab 3 \tab 1 \tab 1 \tab 7 \tab 8 \tab 9 \tab 20 \tab 2 \tab 3544 \cr
}
}

\keyword{variable}
