% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_575.R
\name{racesee}
\alias{racesee}
\title{Observed race}
\source{
General Social Survey https://gss.norc.org
}
\description{
racesee
}
\details{
Question A. CODE WITHOUT ASKING FOR ALL RESPONDENTS EVEN IF UNCERTAIN.
}
\section{Values}{

\itemize{
\item \code{1} white
\item \code{2} black
\item \code{3} other
\item \code{4} hispanic
\item \code{5} black and white
\item \code{6} spanish
\item \code{7} filipino
\item \code{8} american indian
\item \code{9} asian
\item \code{10} yellow
\item \code{11} mexican
\item \code{12} brown
\item \code{13} latino
\item \code{14} middle eastern
\item \code{15} puerto rican
\item \code{16} indian
\item \code{17} alaska native
\item \code{18} pacific islander
\item \code{19} cuban
\item \code{20} japanese
\item \code{21} oriental
\item \code{22} trinidad
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5273/vshow}{GSS Data Explorer page for \code{racesee}}.
}

\seealso{
Other Ethnicity: 
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{eth1}},
\code{\link{eth2}},
\code{\link{eth3}},
\code{\link{ethnic}},
\code{\link{ethnum}},
\code{\link{fucitzn}},
\code{\link{hispanic}},
\code{\link{racecen1}},
\code{\link{racecen1_0022}},
\code{\link{racecen2}},
\code{\link{racecen2_0022}},
\code{\link{racecen3}},
\code{\link{racecen3_0022}},
\code{\link{racedbtf}},
\code{\link{raceself}},
\code{\link{speth1}},
\code{\link{speth2}},
\code{\link{speth3}},
\code{\link{spethnic}},
\code{\link{spethnum}},
\code{\link{uscitzn}},
\code{\link{uscitzn_0822}}
}
\concept{Ethnicity}
\keyword{variable}
