% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_575.R
\name{racesee}
\alias{racesee}
\title{Observed race}
\source{
General Social Survey https://gss.norc.org
}
\description{
racesee
}
\details{
Question A. CODE WITHOUT ASKING FOR ALL RESPONDENTS EVEN IF UNCERTAIN.
}
\section{Values}{

\itemize{
\item \verb{[1]} white
\item \verb{[2]} black
\item \verb{[3]} other
\item \verb{[4]} hispanic
\item \verb{[5]} black and white
\item \verb{[6]} spanish
\item \verb{[7]} filipino
\item \verb{[8]} american indian
\item \verb{[9]} asian
\item \verb{[10]} yellow
\item \verb{[11]} mexican
\item \verb{[12]} brown
\item \verb{[13]} latino
\item \verb{[14]} middle eastern
\item \verb{[15]} puerto rican
\item \verb{[16]} indian
\item \verb{[17]} alaska native
\item \verb{[18]} pacific islander
\item \verb{[19]} cuban
\item \verb{[20]} japanese
\item \verb{[21]} oriental
\item \verb{[22]} trinidad
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5273/vshow}{GSS Data Explorer page for \code{racesee}}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab iap \tab white \tab black \tab other \tab no answer \tab hispanic \tab black and white \tab spanish \tab filipino \tab american indian \tab asian \tab yellow \tab mexican \tab brown \tab latino \tab middle eastern \tab puerto rican \tab indian \tab alaska native \tab pacific islander \tab cuban \tab japanese \tab oriental \tab trinidad \tab Total \cr
   1996 \tab 1444 \tab 1185 \tab 203 \tab 66 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   2000 \tab - \tab 2238 \tab 432 \tab 13 \tab - \tab 36 \tab 3 \tab 8 \tab 8 \tab 13 \tab 42 \tab 1 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 7 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2817 \cr
   Total \tab 1444 \tab 3423 \tab 635 \tab 79 \tab 6 \tab 36 \tab 3 \tab 8 \tab 8 \tab 13 \tab 42 \tab 1 \tab 5 \tab 1 \tab 2 \tab 1 \tab 1 \tab 7 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 5721 \cr
}
}

\seealso{
Other Ethnicity: 
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{hispanic}},
\code{\link{racecen1}},
\code{\link{racecen2}},
\code{\link{racecen3}},
\code{\link{racedbtf}},
\code{\link{raceself}}
}
\concept{Ethnicity}
\keyword{variable}
