% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_575.R
\name{raceself}
\alias{raceself}
\title{R's own racial id}
\source{
General Social Survey https://gss.norc.org
}
\description{
raceself
}
\details{
Question C. What race do you consider yourself?  RECORD VERBATIM AND CODE.
}
\section{Values}{

\itemize{
\item \code{1} white
\item \code{2} black
\item \code{3} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5275/vshow}{GSS Data Explorer page for \code{raceself}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab white \tab black \tab other \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 1139 \tab 200 \tab 117 \tab 1 \tab 3 \tab 2904 \cr
   2000 \tab 1398 \tab 1072 \tab 212 \tab 135 \tab - \tab - \tab 2817 \cr
   Total \tab 2842 \tab 2211 \tab 412 \tab 252 \tab 1 \tab 3 \tab 5721 \cr
}
}

\seealso{
Other Ethnicity: 
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{hispanic}},
\code{\link{racecen1}},
\code{\link{racecen2}},
\code{\link{racecen3}},
\code{\link{racedbtf}},
\code{\link{racesee}}
}
\concept{Ethnicity}
\keyword{variable}
