% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_636.R
\name{racesurv17}
\alias{racesurv17}
\title{How much, if at all, do you think the legacy of slavery affects the position of}
\source{
General Social Survey https://gss.norc.org
}
\description{
racesurv17
}
\details{
Question How much, if at all, do you think the legacy of slavery affects the position of black people in American society today?
}
\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} a fair amount
\item \code{3} not much
\item \code{4} not at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7577/vshow}{GSS Data Explorer page for \code{racesurv17}}.
}

\keyword{variable}
