% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_576.R
\name{racethmn}
\alias{racethmn}
\title{Specify other in racethot}
\source{
General Social Survey https://gss.norc.org
}
\description{
racethmn
}
\details{
Question In describing our background, we often ask someone which racial or ethnic groups best describes them. Imagine if we used instead a 10 point system where points are allocated to whichever racial or ethnic groups the respondent thinks accurately describes themselves. For example, if you think of yourself as just Latino, you would allocate all 10 points to Latino. Or if you think of yourself as mixed race, but more African American than white, you might allocate 6 or 7 points to black and 3 or 4 points to white. Thinking of your background in racial and ethnic terms ... How would you describe your race and ethnicity using this 10-point system using the following choices: G. Specified other ethnicities
}
\section{Values}{

\itemize{
\item \code{1} european
\item \code{2} hispanic
\item \code{3} african
\item \code{4} east asian
\item \code{5} native american
\item \code{6} south asian
\item \code{7} asian, not south or east
\item \code{8} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6649/vshow}{GSS Data Explorer page for \code{racethmn}}.
}

\keyword{variable}
