% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_576.R
\name{racethwh}
\alias{racethwh}
\title{Points assigned to white}
\source{
General Social Survey https://gss.norc.org
}
\description{
racethwh
}
\details{
Question a.  White?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6351/vshow}{GSS Data Explorer page for \code{racethwh}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1888 \tab 259 \tab 16 \tab 545 \tab 16 \tab 9 \tab 8 \tab 35 \tab 12 \tab 13 \tab 26 \tab 22 \tab 3 \tab 15 \tab 2867 \cr
   Total \tab 1888 \tab 259 \tab 16 \tab 545 \tab 16 \tab 9 \tab 8 \tab 35 \tab 12 \tab 13 \tab 26 \tab 22 \tab 3 \tab 15 \tab 2867 \cr
}
}

\seealso{
Other Racial/Ethnic Identity: 
\code{\link{racethas}},
\code{\link{racethbl}},
\code{\link{racethhi}},
\code{\link{racethna}},
\code{\link{racethot}}
}
\concept{Racial/Ethnic Identity}
\keyword{variable}
