% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_48.R
\name{racfirm}
\alias{racfirm}
\title{How firm is r's opinion on racial issues}
\source{
General Social Survey https://gss.norc.org
}
\description{
racfirm
}
\details{
Question C. How firm are you about your opinion on race relations--would you say you are very likely to change your opinion, somewhat likely to change, somewhat unlikely to change, or very unlikely to change?
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely to chang
\item \verb{[2]} somewhat likely
\item \verb{[3]} somewhat unlikely
\item \verb{[4]} very unlikely to cha
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/417/vshow}{GSS Data Explorer page for \code{racfirm}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely to chang \tab somewhat likely \tab somewhat unlikely \tab very unlikely to cha \tab don't know \tab no answer \tab Total \cr
   1983 \tab 795 \tab 31 \tab 227 \tab 270 \tab 246 \tab 27 \tab 3 \tab 1599 \cr
   Total \tab 795 \tab 31 \tab 227 \tab 270 \tab 246 \tab 27 \tab 3 \tab 1599 \cr
}
}

\seealso{
Other Core: 
\code{\link{affrmact}},
\code{\link{blksimp}},
\code{\link{raccare}},
\code{\link{racgive}},
\code{\link{racinfo}},
\code{\link{racjoin}},
\code{\link{racthink}},
\code{\link{racwrite}},
\code{\link{wrkwayup}}
}
\concept{Core}
\keyword{variable}
