% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_48.R
\name{racgive}
\alias{racgive}
\title{Has r given money to racial orgs}
\source{
General Social Survey https://gss.norc.org
}
\description{
racgive
}
\details{
Question Have you ever . . .
B. Given money to an organization concerned with this issue?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1983 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/421/vshow}{GSS Data Explorer page for \code{racgive}}.
}

\seealso{
Other Core: 
\code{\link{affrmact}},
\code{\link{blksimp}},
\code{\link{raccare}},
\code{\link{racfirm}},
\code{\link{racinfo}},
\code{\link{racjoin}},
\code{\link{racthink}},
\code{\link{racwrite}},
\code{\link{wrkwayup}}
}
\concept{Core}
\keyword{variable}
