% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_46.R
\name{rachaf}
\alias{rachaf}
\title{Children to school half opposite race}
\source{
General Social Survey https://gss.norc.org
}
\description{
rachaf
}
\details{
Question B. IF NO OR DON''T KNOW TO A:  Where half of the children are (Whites/(Negroes/Blacks/African- Americans)?
}
\section{Values}{

\itemize{
\item \code{1} yes, object
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/- \tab full \cr
   1974 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1988 \tab A/B/- \tab partial \cr
   1989 \tab A/B/- \tab partial \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/- \tab partial \cr
   1993 \tab A/B/- \tab partial \cr
   1994 \tab A/B/- \tab partial \cr
   1996 \tab A/B/- \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/396/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllll}{
   year \tab don't know \tab iap \tab no \tab no answer \tab yes, object \tab not available in this year \tab Total \cr
   1972 \tab 38 \tab 357 \tab 997 \tab 5 \tab 216 \tab - \tab 1613 \cr
   1973 \tab - \tab 1504 \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 38 \tab 240 \tab 889 \tab 19 \tab 298 \tab - \tab 1484 \cr
   1975 \tab 48 \tab 253 \tab 912 \tab 6 \tab 271 \tab - \tab 1490 \cr
   1976 \tab - \tab 1499 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 23 \tab 274 \tab 990 \tab 11 \tab 232 \tab - \tab 1530 \cr
   1978 \tab 42 \tab 70 \tab 1156 \tab 7 \tab 257 \tab - \tab 1532 \cr
   1980 \tab - \tab 1468 \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 44 \tab 104 \tab 1447 \tab 9 \tab 256 \tab - \tab 1860 \cr
   1983 \tab 37 \tab 78 \tab 1209 \tab 8 \tab 267 \tab - \tab 1599 \cr
   1984 \tab - \tab 1473 \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 29 \tab 64 \tab 1195 \tab 9 \tab 237 \tab - \tab 1534 \cr
   1986 \tab 33 \tab 70 \tab 1113 \tab 8 \tab 246 \tab - \tab 1470 \cr
   1987 \tab - \tab 1819 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 22 \tab 534 \tab 773 \tab 4 \tab 148 \tab - \tab 1481 \cr
   1989 \tab 36 \tab 573 \tab 757 \tab 8 \tab 163 \tab - \tab 1537 \cr
   1990 \tab 37 \tab 42 \tab 1074 \tab 6 \tab 213 \tab - \tab 1372 \cr
   1991 \tab 22 \tab 526 \tab 813 \tab 10 \tab 146 \tab - \tab 1517 \cr
   1993 \tab 31 \tab 573 \tab 850 \tab 4 \tab 148 \tab - \tab 1606 \cr
   1994 \tab 45 \tab 1086 \tab 1616 \tab 15 \tab 230 \tab - \tab 2992 \cr
   1996 \tab 15 \tab 1942 \tab 793 \tab 14 \tab 140 \tab - \tab 2904 \cr
   1998 \tab - \tab 2832 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab 2817 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 2765 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 2812 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 4510 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 2023 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 2044 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 1974 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 2538 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 2867 \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 2348 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 4032 \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 540 \tab 48111 \tab 16584 \tab 143 \tab 3468 \tab 6853 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{color}},
\code{\link{racchurh}},
\code{\link{rachisch}},
\code{\link{racjob}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racpres}}

Other Race: 
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{color}},
\code{\link{racchurh}},
\code{\link{rachisch}},
\code{\link{racjob}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racpres}}
}
\concept{Core}
\concept{Race}
\keyword{variable}
