% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_46.R
\name{rachisch}
\alias{rachisch}
\title{How integrated was r's high school}
\source{
General Social Survey https://gss.norc.org
}
\description{
rachisch
}
\details{
Question Was the high school you attended all black, integrated--but mostly black, or integrated--but mostly white?
}
\section{Values}{

\itemize{
\item \code{0} didnt go to hs
\item \code{1} all black
\item \code{2} mostly black
\item \code{3} mostly white
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/405/vshow}{GSS Data Explorer page for \code{rachisch}}.
}

\seealso{
Other Core: 
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{color}},
\code{\link{racchurh}},
\code{\link{rachaf}},
\code{\link{racjob}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racpres}}

Other Race: 
\code{\link{busing}},
\code{\link{busing10}},
\code{\link{color}},
\code{\link{racchurh}},
\code{\link{rachaf}},
\code{\link{racjob}},
\code{\link{racmost}},
\code{\link{racname}},
\code{\link{racpres}}
}
\concept{Core}
\concept{Race}
\keyword{variable}
