% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_45.R
\name{rachome}
\alias{rachome}
\title{Opp. race home for dinner recently}
\source{
General Social Survey https://gss.norc.org
}
\description{
rachome
}
\details{
Question During the last few years, has anyone in your family brought a friend who was a (Negro/Black/African-American) home for dinner?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/- \tab full \cr
   1974 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab A/C \tab partial \cr
   1989 \tab A/C \tab partial \cr
   1990 \tab A/C \tab partial \cr
   1991 \tab A/C \tab partial \cr
   1993 \tab A/C \tab partial \cr
   1994 \tab A/C \tab partial \cr
   1996 \tab A/C \tab partial \cr
   2006 \tab D/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/393/vshow}{GSS Data Explorer page for \code{rachome}}.
}

\seealso{
Other Core: 
\code{\link{racclos}},
\code{\link{racdis}},
\code{\link{racfew}},
\code{\link{racinteg}},
\code{\link{raclive}},
\code{\link{racobjct}},
\code{\link{racopen}},
\code{\link{racschol}},
\code{\link{racseg}}
}
\concept{Core}
\keyword{variable}
