% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_47.R
\name{racimp}
\alias{racimp}
\title{Importance of race relations issues to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
racimp
}
\details{
Question A. How important is the race relations issue to you--would you say it is one of the most important, important, not very important, or not important at all?
}
\section{Values}{

\itemize{
\item \code{1} one of most imp
\item \code{2} important
\item \code{3} not very imp
\item \code{4} not imp at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/415/vshow}{GSS Data Explorer page for \code{racimp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab one of most imp \tab important \tab not very imp \tab not imp at all \tab don't know \tab no answer \tab Total \cr
   1983 \tab 795 \tab 76 \tab 342 \tab 286 \tab 86 \tab 12 \tab 2 \tab 1599 \cr
   Total \tab 795 \tab 76 \tab 342 \tab 286 \tab 86 \tab 12 \tab 2 \tab 1599 \cr
}
}

\seealso{
Other Core: 
\code{\link{racinc}},
\code{\link{racmix}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racocc}},
\code{\link{racopnow}},
\code{\link{racopwil}},
\code{\link{racparty}},
\code{\link{ractrust}}
}
\concept{Core}
\keyword{variable}
