% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_47.R
\name{racinc}
\alias{racinc}
\title{Equality in income}
\source{
General Social Survey https://gss.norc.org
}
\description{
racinc
}
\details{
Question If a black person has the same qualifications as a white person, Do you feel that he or she can make as much money . . .
}
\section{Values}{

\itemize{
\item \code{1} almost always
\item \code{2} sometimes
\item \code{3} almost never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/412/vshow}{GSS Data Explorer page for \code{racinc}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab almost always \tab sometimes \tab almost never \tab don't know \tab no answer \tab Total \cr
   1982 \tab 1350 \tab 66 \tab 278 \tab 159 \tab 2 \tab 5 \tab 1860 \cr
   Total \tab 1350 \tab 66 \tab 278 \tab 159 \tab 2 \tab 5 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{racimp}},
\code{\link{racmix}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racocc}},
\code{\link{racopnow}},
\code{\link{racopwil}},
\code{\link{racparty}},
\code{\link{ractrust}}
}
\concept{Core}
\keyword{variable}
