% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_48.R
\name{racinfo}
\alias{racinfo}
\title{How much info does r have on race issues}
\source{
General Social Survey https://gss.norc.org
}
\description{
racinfo
}
\details{
Question B. How much information do you have about the race relations issue. Do you have all the information you need, most of the information, some information, or very little information?
}
\section{Values}{

\itemize{
\item \code{1} all info r needs
\item \code{2} most info
\item \code{3} some info
\item \code{4} very little info
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1983 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/416/vshow}{GSS Data Explorer page for \code{racinfo}}.
}

\seealso{
Other Core: 
\code{\link{affrmact}},
\code{\link{blksimp}},
\code{\link{raccare}},
\code{\link{racfirm}},
\code{\link{racgive}},
\code{\link{racjoin}},
\code{\link{racthink}},
\code{\link{racwrite}},
\code{\link{wrkwayup}}

Other Race: 
\code{\link{blksimp}},
\code{\link{racfirm}},
\code{\link{wrkwayup}}

Other Split Ballots: 
\code{\link{raccare}},
\code{\link{racfirm}},
\code{\link{racgive}},
\code{\link{racjoin}},
\code{\link{racthink}},
\code{\link{racwrite}}
}
\concept{Core}
\concept{Race}
\concept{Split Ballots}
\keyword{variable}
