% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_44.R
\name{racmar10}
\alias{racmar10}
\title{Favor law against rac intrmar 10 yrs ago}
\source{
General Social Survey https://gss.norc.org
}
\description{
racmar10
}
\details{
Question B. Now thinking about ten years ago, that is in 1972, did you then think there should be laws against marriages between (Negroes/Blacks) and whites?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/383/vshow}{GSS Data Explorer page for \code{racmar10}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 585 \tab 1161 \tab 97 \tab 17 \tab 1860 \cr
   Total \tab 0 \tab 585 \tab 1161 \tab 97 \tab 17 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{cathtemp}},
\code{\link{feclergy}},
\code{\link{fepriest}},
\code{\link{jewtemp}},
\code{\link{mslmtemp}},
\code{\link{racdin}},
\code{\link{racmar}},
\code{\link{racpush}},
\code{\link{relgrade}}
}
\concept{Core}
\keyword{variable}
