% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_81.R
\name{racmarel}
\alias{racmarel}
\title{What if relative married other race}
\source{
General Social Survey https://gss.norc.org
}
\description{
racmarel
}
\details{
Question How would it make you feel if a close relative of yours were planning to marry a ((Negro/Black)/White)?  Would you be very uneasy, somewhat uneasy, or not uneasy at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} very uneasy
\item \verb{[2]} somewhat uneasy
\item \verb{[3]} not uneasy
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/721/vshow}{GSS Data Explorer page for \code{racmarel}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very uneasy \tab somewhat uneasy \tab not uneasy \tab don't know \tab no answer \tab Total \cr
   1977 \tab - \tab 635 \tab 481 \tab 370 \tab 38 \tab 6 \tab 1530 \cr
   Total \tab 0 \tab 635 \tab 481 \tab 370 \tab 38 \tab 6 \tab 1530 \cr
}
}

\seealso{
Other Core: 
\code{\link{racavoid}},
\code{\link{racchng}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racmarpr}},
\code{\link{racquit}},
\code{\link{racsups}},
\code{\link{racteach}}
}
\concept{Core}
\keyword{variable}
