% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_47.R
\name{racmix}
\alias{racmix}
\title{How many friends are white}
\source{
General Social Survey https://gss.norc.org
}
\description{
racmix
}
\details{
Question How many of your close friends are white?
}
\section{Values}{

\itemize{
\item \code{1} none
\item \code{2} one or two
\item \code{3} three or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/406/vshow}{GSS Data Explorer page for \code{racmix}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab none \tab one or two \tab three or more \tab no answer \tab Total \cr
   1982 \tab 1350 \tab 170 \tab 124 \tab 214 \tab 2 \tab 1860 \cr
   Total \tab 1350 \tab 170 \tab 124 \tab 214 \tab 2 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racocc}},
\code{\link{racopnow}},
\code{\link{racopwil}},
\code{\link{racparty}},
\code{\link{ractrust}}
}
\concept{Core}
\keyword{variable}
