% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_47.R
\name{racnobuy}
\alias{racnobuy}
\title{Ever boycott certain products}
\source{
General Social Survey https://gss.norc.org
}
\description{
racnobuy
}
\details{
Question Have you ever stopped buying certain products because of the way the company or country which makes them has treated black people?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/408/vshow}{GSS Data Explorer page for \code{racnobuy}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1982 \tab 1350 \tab 135 \tab 369 \tab 6 \tab 1860 \cr
   Total \tab 1350 \tab 135 \tab 369 \tab 6 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racmix}},
\code{\link{racneigh}},
\code{\link{racocc}},
\code{\link{racopnow}},
\code{\link{racopwil}},
\code{\link{racparty}},
\code{\link{ractrust}}
}
\concept{Core}
\keyword{variable}
