% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_45.R
\name{racobjct}
\alias{racobjct}
\title{Neighborhood integrated by same ses}
\source{
General Social Survey https://gss.norc.org
}
\description{
racobjct
}
\details{
Question If a Negro with the same income and education as you have, moved into your block, would it make any difference to you?
}
\section{Values}{

\itemize{
\item \code{1} yes,would like it
\item \code{2} yes,would not like
\item \code{3} no difference
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1972 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/392/vshow}{GSS Data Explorer page for \code{racobjct}}.

Counts by year:\tabular{llllllll}{
   year \tab yes,would like it \tab yes,would not like \tab no difference \tab don't know \tab iap \tab no answer \tab Total \cr
   1972 \tab 17 \tab 186 \tab 1117 \tab 29 \tab 261 \tab 3 \tab 1613 \cr
   Total \tab 17 \tab 186 \tab 1117 \tab 29 \tab 261 \tab 3 \tab 1613 \cr
}
}

\seealso{
Other Core: 
\code{\link{racclos}},
\code{\link{racdis}},
\code{\link{racfew}},
\code{\link{rachome}},
\code{\link{racinteg}},
\code{\link{raclive}},
\code{\link{racopen}},
\code{\link{racschol}},
\code{\link{racseg}}
}
\concept{Core}
\keyword{variable}
