% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_639.R
\name{racopennv}
\alias{racopennv}
\title{Vote on open housing law (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
racopennv
}
\details{
Question Suppose there is a community-wide vote on the general housing issue. There are two possible laws to vote on: One law says that a homeowner can decide whom to sell his house to, even if that homeowner prefers not to sell to someone because of their race or color. The second law says that a homeowner cannot refuse to sell to someone because of their race or color. Which law would you vote for?
}
\section{Values}{

\itemize{
\item \code{1} first law
\item \code{2} second law
\item \code{3} neither (phone mode only: volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021 \cr
   A/C/- \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7608/vshow}{GSS Data Explorer page for \code{racopennv}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab first law \tab second law \tab neither (phone mode only: volunteered) \tab don't know \tab skipped on web \tab no answer \tab Total \cr
   2021 \tab 2045 \tab 353 \tab 1600 \tab 9 \tab 2 \tab 23 \tab - \tab 4032 \cr
   2022 \tab 2981 \tab 100 \tab 454 \tab - \tab - \tab 8 \tab 1 \tab 3544 \cr
   Total \tab 5026 \tab 453 \tab 2054 \tab 9 \tab 2 \tab 31 \tab 1 \tab 7576 \cr
}
}

\keyword{variable}
