% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_639.R
\name{racopennv}
\alias{racopennv}
\title{Vote on open housing law (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
racopennv
}
\details{
Question Suppose there is a community-wide vote on the general housing issue. There are two possible laws to vote on: One law says that a homeowner can decide whom to sell his house to, even if that homeowner prefers not to sell to someone because of their race or color. The second law says that a homeowner cannot refuse to sell to someone because of their race or color. Which law would you vote for?
}
\section{Values}{

\itemize{
\item \code{1} first law
\item \code{2} second law
\item \code{3} neither (phone mode only: volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/C/- \tab partial \cr
   2024 \tab A/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7608/vshow}{GSS Data Explorer page for \code{racopennv}}.
}

\keyword{variable}
