% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_47.R
\name{racopnow}
\alias{racopnow}
\title{Opportunities improved in last 5 yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
racopnow
}
\details{
Question A. Do you think the opportunities for blacks to get ahead have improved in the last five years, remained about the same, or gotten worse?
}
\section{Values}{

\itemize{
\item \code{1} improved
\item \code{2} remained same
\item \code{3} gotten worse
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/413/vshow}{GSS Data Explorer page for \code{racopnow}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab improved \tab remained same \tab gotten worse \tab don't know \tab no answer \tab Total \cr
   1982 \tab 1350 \tab 241 \tab 161 \tab 100 \tab 4 \tab 4 \tab 1860 \cr
   Total \tab 1350 \tab 241 \tab 161 \tab 100 \tab 4 \tab 4 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racmix}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racocc}},
\code{\link{racopwil}},
\code{\link{racparty}},
\code{\link{ractrust}}
}
\concept{Core}
\keyword{variable}
