% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_47.R
\name{racopwil}
\alias{racopwil}
\title{Opportunities improve in next 5 yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
racopwil
}
\details{
Question B. In the next five years, do you think that opportunities for blacks to get ahead will improve, remain about the same, or get worse?
}
\section{Values}{

\itemize{
\item \code{1} improve
\item \code{2} remain same
\item \code{3} get worse
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/414/vshow}{GSS Data Explorer page for \code{racopwil}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab improve \tab remain same \tab get worse \tab don't know \tab no answer \tab Total \cr
   1982 \tab 1350 \tab 177 \tab 168 \tab 145 \tab 14 \tab 6 \tab 1860 \cr
   Total \tab 1350 \tab 177 \tab 168 \tab 145 \tab 14 \tab 6 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racmix}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racocc}},
\code{\link{racopnow}},
\code{\link{racparty}},
\code{\link{ractrust}}
}
\concept{Core}
\keyword{variable}
