% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_139.R
\name{racpromo}
\alias{racpromo}
\title{Ethnic background makes your promotion different}
\source{
General Social Survey https://gss.norc.org
}
\description{
racpromo
}
\details{
Question Do you think your race or ethnic background makes your promotion opportunities better or worse?
}
\section{Values}{

\itemize{
\item \code{1} better
\item \code{2} worse
\item \code{3} no effect
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1305/vshow}{GSS Data Explorer page for \code{racpromo}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab better \tab worse \tab no effect \tab don't know \tab no answer \tab Total \cr
   1991 \tab 605 \tab 204 \tab 92 \tab 584 \tab 17 \tab 15 \tab 1517 \cr
   Total \tab 605 \tab 204 \tab 92 \tab 584 \tab 17 \tab 15 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{advances}},
\code{\link{futpromo}},
\code{\link{imppromo}},
\code{\link{jobcntrl}},
\code{\link{lottosay}},
\code{\link{numpromo}},
\code{\link{promoted}},
\code{\link{sexpromo}},
\code{\link{wrkindep}}
}
\concept{Work Organizations}
\keyword{variable}
