% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_44.R
\name{racpush}
\alias{racpush}
\title{Blacks shouldn't push}
\source{
General Social Survey https://gss.norc.org
}
\description{
racpush
}
\details{
Question Here are some opinions other people have expressed in connection with (Negro/Black)-white relations. Which statement on the card comes closest to how you, yourself, feel? The first one is . . .
A. (Negroes/Blacks/African-Americans) shouldn't push themselves where they're not wanted.
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree slightly
\item \code{3} disagree slightly
\item \code{4} disagree strongly
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/- \tab full \cr
   1973 \tab -/- \tab full \cr
   1975 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1994 \tab A/C \tab partial \cr
   1996 \tab A/C \tab partial \cr
   1998 \tab A/C \tab partial \cr
   2000 \tab A/C \tab partial \cr
   2002 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/385/vshow}{GSS Data Explorer page for \code{racpush}}.
}

\seealso{
Other Core: 
\code{\link{cathtemp}},
\code{\link{feclergy}},
\code{\link{fepriest}},
\code{\link{jewtemp}},
\code{\link{mslmtemp}},
\code{\link{racdin}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{relgrade}}

Other Race: 
\code{\link{racdin}},
\code{\link{racmar}},
\code{\link{racmar10}}
}
\concept{Core}
\concept{Race}
\keyword{variable}
