% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_81.R
\name{racquit}
\alias{racquit}
\title{Quit if couldn't integrate club}
\source{
General Social Survey https://gss.norc.org
}
\description{
racquit
}
\details{
Question If you and your friends belonged to a social club that would not let (Whites/Negroes/Blacks) join, would you try to change the rules so that (Whites/(Negroes/Blacks/African-Americans) could join?
A. IF YES:  If you could not get the rules changed, do you think you would resign from the club, even if your friends didn't?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1977 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/727/vshow}{GSS Data Explorer page for \code{racquit}}.
}

\seealso{
Other Core: 
\code{\link{racavoid}},
\code{\link{racchng}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racsups}},
\code{\link{racteach}}

Other Race: 
\code{\link{racavoid}},
\code{\link{racchng}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racsups}},
\code{\link{racteach}}
}
\concept{Core}
\concept{Race}
\keyword{variable}
