% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_45.R
\name{racschol}
\alias{racschol}
\title{Should whites & blacks go to same school}
\source{
General Social Survey https://gss.norc.org
}
\description{
racschol
}
\details{
Question Was the high school you attended all black, integrated - but mostly black, or integrated - but mostly white?
}
\section{Values}{

\itemize{
\item \code{1} same schools
\item \code{2} separate schools
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab - \tab full \cr
   1976 \tab - \tab full \cr
   1977 \tab - \tab full \cr
   1980 \tab - \tab full \cr
   1982 \tab - \tab full \cr
   1984 \tab - \tab full \cr
   1985 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/394/vshow}{GSS Data Explorer page for \code{racschol}}.
}

\seealso{
Other Core: 
\code{\link{racclos}},
\code{\link{racdis}},
\code{\link{racfew}},
\code{\link{rachome}},
\code{\link{racinteg}},
\code{\link{raclive}},
\code{\link{racobjct}},
\code{\link{racopen}},
\code{\link{racseg}}
}
\concept{Core}
\keyword{variable}
