% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_435.R
\name{racspeak}
\alias{racspeak}
\title{Allow racists to hold public meetings}
\source{
General Social Survey https://gss.norc.org
}
\description{
racspeak
}
\details{
Question Second, consider people who believe that whites are racially superior to all other races.
Circle a number after each question to show if you think such people should be allowed to . . .
A. Hold public meetings to express their views.
}
\section{Values}{

\itemize{
\item \code{1} definitely allowed
\item \code{2} probably allowed
\item \code{3} probably not allowed
\item \code{4} definitely not allowed
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3991/vshow}{GSS Data Explorer page for \code{racspeak}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely allowed \tab probably allowed \tab probably not allowed \tab definitely not allowed \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 185 \tab 175 \tab 101 \tab 174 \tab 34 \tab 8 \tab 1534 \cr
   1990 \tab 155 \tab 308 \tab 387 \tab 220 \tab 217 \tab 70 \tab 15 \tab 1372 \cr
   Total \tab 1012 \tab 493 \tab 562 \tab 321 \tab 391 \tab 104 \tab 23 \tab 2906 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{crimhold}},
\code{\link{crimread}},
\code{\link{crimtail}},
\code{\link{crimtap}},
\code{\link{racpub}},
\code{\link{ractch15}},
\code{\link{revpub}},
\code{\link{revspeak}},
\code{\link{revtch15}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
