% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_80.R
\name{racsubgv}
\alias{racsubgv}
\title{Suburban govts encourage integration}
\source{
General Social Survey https://gss.norc.org
}
\description{
racsubgv
}
\details{
Question What about the city government in white suburbs? Do you think they should encourage (Negro/Black) people to buy homes in the suburbs, should they discourage them, or should they leave it to private efforts?
}
\section{Values}{

\itemize{
\item \code{1} encourage
\item \code{2} discourage
\item \code{3} leave alone
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/720/vshow}{GSS Data Explorer page for \code{racsubgv}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab encourage \tab discourage \tab leave alone \tab don't know \tab no answer \tab Total \cr
   1977 \tab - \tab 193 \tab 146 \tab 1130 \tab 55 \tab 6 \tab 1530 \cr
   Total \tab 0 \tab 193 \tab 146 \tab 1130 \tab 55 \tab 6 \tab 1530 \cr
}
}

\seealso{
Other Core: 
\code{\link{erafirm}},
\code{\link{eraimp}},
\code{\link{erainfo}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}},
\code{\link{febear}},
\code{\link{feworkif}},
\code{\link{racsubs}}
}
\concept{Core}
\keyword{variable}
