% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_80.R
\name{racsubs}
\alias{racsubs}
\title{Favor attempts to integrate suburbs}
\source{
General Social Survey https://gss.norc.org
}
\description{
racsubs
}
\details{
Question Some religious and business groups have set up programs to encourage (Negro/Black) people to buy houses in white suburbs. Do you favor or oppose these voluntary programs to integrate white suburbs?
}
\section{Values}{

\itemize{
\item \code{1} favor
\item \code{2} oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/719/vshow}{GSS Data Explorer page for \code{racsubs}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab favor \tab oppose \tab don't know \tab no answer \tab Total \cr
   1977 \tab - \tab 585 \tab 832 \tab 106 \tab 7 \tab 1530 \cr
   Total \tab 0 \tab 585 \tab 832 \tab 106 \tab 7 \tab 1530 \cr
}
}

\seealso{
Other Core: 
\code{\link{erafirm}},
\code{\link{eraimp}},
\code{\link{erainfo}},
\code{\link{erawhy1}},
\code{\link{erawhy2}},
\code{\link{erawhy3}},
\code{\link{febear}},
\code{\link{feworkif}},
\code{\link{racsubgv}}
}
\concept{Core}
\keyword{variable}
