% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_81.R
\name{racsups}
\alias{racsups}
\title{Expect probs if blacks supervise whites}
\source{
General Social Survey https://gss.norc.org
}
\description{
racsups
}
\details{
Question Please tell me whether you agree or disagree with the following statements . . .
B. You can expect special problems with (Negroes/Black) supervisors getting along with workers that are mostly White.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree
\item \verb{[2]} disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/723/vshow}{GSS Data Explorer page for \code{racsups}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab no answer \tab Total \cr
   1977 \tab - \tab 714 \tab 734 \tab 76 \tab 6 \tab 1530 \cr
   Total \tab 0 \tab 714 \tab 734 \tab 76 \tab 6 \tab 1530 \cr
}
}

\seealso{
Other Core: 
\code{\link{racavoid}},
\code{\link{racchng}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racquit}},
\code{\link{racteach}}
}
\concept{Core}
\keyword{variable}
