% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_435.R
\name{ractch15}
\alias{ractch15}
\title{Allow racists to teach 15 yr. olds}
\source{
General Social Survey https://gss.norc.org
}
\description{
ractch15
}
\details{
Question Second, consider people who believe that whites are racially superior to all other races.
Circle a number after each question to show if you think such people should be allowed to . . .
B. Teach 15 year olds in schools.
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely allowed
\item \verb{[2]} probably allowed
\item \verb{[3]} probably not allowed
\item \verb{[4]} definitely not allowed
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3992/vshow}{GSS Data Explorer page for \code{ractch15}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely allowed \tab probably allowed \tab probably not allowed \tab definitely not allowed \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 59 \tab 85 \tab 143 \tab 333 \tab 43 \tab 14 \tab 1534 \cr
   Total \tab 857 \tab 59 \tab 85 \tab 143 \tab 333 \tab 43 \tab 14 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{crimhold}},
\code{\link{crimread}},
\code{\link{crimtail}},
\code{\link{crimtap}},
\code{\link{racpub}},
\code{\link{racspeak}},
\code{\link{revpub}},
\code{\link{revspeak}},
\code{\link{revtch15}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
