% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_81.R
\name{racteach}
\alias{racteach}
\title{Don't hire teachers against integration}
\source{
General Social Survey https://gss.norc.org
}
\description{
racteach
}
\details{
Question Please tell me whether you agree or disagree with the following statements . . .
C. A school board should not hire a person to teach if that person belongs to an organization that opposes school integration.
}
\section{Values}{

\itemize{
\item \code{1} agree
\item \code{2} disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/724/vshow}{GSS Data Explorer page for \code{racteach}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab agree \tab disagree \tab don't know \tab no answer \tab Total \cr
   1977 \tab - \tab 707 \tab 710 \tab 103 \tab 10 \tab 1530 \cr
   Total \tab 0 \tab 707 \tab 710 \tab 103 \tab 10 \tab 1530 \cr
}
}

\seealso{
Other Core: 
\code{\link{racavoid}},
\code{\link{racchng}},
\code{\link{racdif1}},
\code{\link{racdif2}},
\code{\link{racdif3}},
\code{\link{racmarel}},
\code{\link{racmarpr}},
\code{\link{racquit}},
\code{\link{racsups}}
}
\concept{Core}
\keyword{variable}
