% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_48.R
\name{racthink}
\alias{racthink}
\title{How often r & friends think about race issues}
\source{
General Social Survey https://gss.norc.org
}
\description{
racthink
}
\details{
Question How often would you say that you and your friends think about race relations?  Would you say that you and your friends think about race relations very often, fairly often, occasionally, or almost never?
}
\section{Values}{

\itemize{
\item \verb{[1]} very often
\item \verb{[2]} fairly often
\item \verb{[3]} occasionally
\item \verb{[4]} almost never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/419/vshow}{GSS Data Explorer page for \code{racthink}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very often \tab fairly often \tab occasionally \tab almost never \tab don't know \tab no answer \tab Total \cr
   1983 \tab 804 \tab 56 \tab 109 \tab 346 \tab 274 \tab 4 \tab 6 \tab 1599 \cr
   Total \tab 804 \tab 56 \tab 109 \tab 346 \tab 274 \tab 4 \tab 6 \tab 1599 \cr
}
}

\seealso{
Other Core: 
\code{\link{affrmact}},
\code{\link{blksimp}},
\code{\link{raccare}},
\code{\link{racfirm}},
\code{\link{racgive}},
\code{\link{racinfo}},
\code{\link{racjoin}},
\code{\link{racwrite}},
\code{\link{wrkwayup}}
}
\concept{Core}
\keyword{variable}
