% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_47.R
\name{ractrust}
\alias{ractrust}
\title{Trust white people}
\source{
General Social Survey https://gss.norc.org
}
\description{
ractrust
}
\details{
Question Do you feel you can trust most white people, some white people, or no white people?
}
\section{Values}{

\itemize{
\item \verb{[1]} most white people
\item \verb{[2]} some white people
\item \verb{[3]} no white people
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/409/vshow}{GSS Data Explorer page for \code{ractrust}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab most white people \tab some white people \tab no white people \tab no answer \tab Total \cr
   1982 \tab 1350 \tab 36 \tab 411 \tab 55 \tab 8 \tab 1860 \cr
   Total \tab 1350 \tab 36 \tab 411 \tab 55 \tab 8 \tab 1860 \cr
}
}

\seealso{
Other Core: 
\code{\link{racimp}},
\code{\link{racinc}},
\code{\link{racmix}},
\code{\link{racneigh}},
\code{\link{racnobuy}},
\code{\link{racocc}},
\code{\link{racopnow}},
\code{\link{racopwil}},
\code{\link{racparty}}
}
\concept{Core}
\keyword{variable}
