% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_133.R
\name{racwork}
\alias{racwork}
\title{Racial makeup of workplace}
\source{
General Social Survey https://gss.norc.org
}
\description{
racwork
}
\details{
Question Are the people who work where you work all white, mostly white, about half and half, mostly black, or all black?
}
\section{Values}{

\itemize{
\item \code{1} all white
\item \code{2} mostly white
\item \code{3} about half and half
\item \code{4} mostly black
\item \code{5} all black
\item \code{6} works alone
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
   1996 \tab A/C/- \tab partial \cr
   1998 \tab A/C/- \tab partial \cr
   2000 \tab A/C/- \tab partial \cr
   2002 \tab A/C/- \tab partial \cr
   2004 \tab A/C/- \tab partial \cr
   2006 \tab A/C/- \tab partial \cr
   2008 \tab A/C/- \tab partial \cr
   2010 \tab A/C/- \tab partial \cr
   2012 \tab A/C/- \tab partial \cr
   2014 \tab A/C/- \tab partial \cr
   2016 \tab A/C/- \tab partial \cr
   2018 \tab A/C/- \tab partial \cr
   2021 \tab A/C/- \tab partial \cr
   2022 \tab A/C/- \tab partial \cr
   2024 \tab A/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1249/vshow}{GSS Data Explorer page for \code{racwork}}.
}

\seealso{
Other Intergroup Relations: 
\code{\link{asnhouse}},
\code{\link{asnjobs}},
\code{\link{blkhouse}},
\code{\link{blkjobs}},
\code{\link{hsphouse}},
\code{\link{hspjobs}},
\code{\link{influasn}},
\code{\link{influhsp}},
\code{\link{influso}}
}
\concept{Intergroup Relations}
\keyword{variable}
