% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_48.R
\name{racwrite}
\alias{racwrite}
\title{R ever written official abt race issues}
\source{
General Social Survey https://gss.norc.org
}
\description{
racwrite
}
\details{
Question Have you ever . . .
A. Written a letter to a public official expressing your views on race relations?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1983 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/420/vshow}{GSS Data Explorer page for \code{racwrite}}.
}

\seealso{
Other Core: 
\code{\link{affrmact}},
\code{\link{blksimp}},
\code{\link{raccare}},
\code{\link{racfirm}},
\code{\link{racgive}},
\code{\link{racinfo}},
\code{\link{racjoin}},
\code{\link{racthink}},
\code{\link{wrkwayup}}

Other Split Ballots: 
\code{\link{raccare}},
\code{\link{racfirm}},
\code{\link{racgive}},
\code{\link{racinfo}},
\code{\link{racjoin}},
\code{\link{racthink}}
}
\concept{Core}
\concept{Split Ballots}
\keyword{variable}
