% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_78.R
\name{radiohrs}
\alias{radiohrs}
\title{Hours per day listening to radio}
\source{
General Social Survey https://gss.norc.org
}
\description{
radiohrs
}
\details{
Question Do you ever listen to the radio?
Yes…..(ASK RADIOHRS)
No.....
A. IF YES: On the average, about how many hours a day do you usually listen to the radio?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1978 \tab - \tab full \cr
   1982 \tab - \tab full \cr
   1983 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/692/vshow}{GSS Data Explorer page for \code{radiohrs}}.
}

\seealso{
Other Core: 
\code{\link{abhave1}},
\code{\link{abhave2}},
\code{\link{abspno}},
\code{\link{comprend}},
\code{\link{coop}},
\code{\link{coop2}},
\code{\link{form}},
\code{\link{phone}}
}
\concept{Core}
\concept{Radio Listening}
\keyword{variable}
