% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_63.R
\name{rank}
\alias{rank}
\title{R's self ranking of social position}
\source{
General Social Survey https://gss.norc.org
}
\description{
rank
}
\details{
Question In our society there are groups which tend to be towards the top and those that are towards the bottom. Here we have a scale that runs from top to bottom.
A. Where would you put yourself on this scale?
HAND SCALE AND PENCIL TO RESPONDENT. LET RESPONDENT MARK SCALE. Be sure that the mark is within one of the boxes. Record answer below.
}
\section{Values}{

\itemize{
\item \code{1} 1 - top
\item \code{10} 10 - bottom
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1983 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/570/vshow}{GSS Data Explorer page for \code{rank}}.
}

\seealso{
Other Core: 
\code{\link{class}()},
\code{\link{classy}},
\code{\link{jobaccmp}},
\code{\link{jobhelp}},
\code{\link{jobpeop}},
\code{\link{jobresp}},
\code{\link{jobsafe}},
\code{\link{jobsoc}},
\code{\link{rank10}}
}
\concept{Core}
\concept{ISSP Social Inequality}
\keyword{variable}
