% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_63.R
\name{rank10}
\alias{rank10}
\title{R's social rank 10 years ago}
\source{
General Social Survey https://gss.norc.org
}
\description{
rank10
}
\details{
Question In our society there are groups which tend to be towards the top and those that are towards the bottom. Here we have a scale that runs from top to bottom.
B. And ten years ago, where did you fit in then?
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 - top
\item \verb{[10]} 10 - bottom
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/571/vshow}{GSS Data Explorer page for \code{rank10}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 1 - top \tab 10 - bottom \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab can't choose \tab no answer \tab Total \cr
   2000 \tab 1545 \tab 27 \tab 40 \tab 49 \tab 106 \tab 132 \tab 239 \tab 194 \tab 149 \tab 133 \tab 52 \tab 7 \tab 144 \tab 2817 \cr
   Total \tab 1545 \tab 27 \tab 40 \tab 49 \tab 106 \tab 132 \tab 239 \tab 194 \tab 149 \tab 133 \tab 52 \tab 7 \tab 144 \tab 2817 \cr
}
}

\seealso{
Other Core: 
\code{\link{class}()},
\code{\link{classy}},
\code{\link{jobaccmp}},
\code{\link{jobhelp}},
\code{\link{jobpeop}},
\code{\link{jobresp}},
\code{\link{jobsafe}},
\code{\link{jobsoc}},
\code{\link{rank}()}
}
\concept{Core}
\keyword{variable}
