% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_631.R
\name{rank10fut}
\alias{rank10fut}
\title{R's social rank 10 years later}
\source{
General Social Survey https://gss.norc.org
}
\description{
rank10fut
}
\details{
Question And thinking ahead 10 years from now, where do you think you will be on a scale of 1 to 10, where 1 is the top and 10 the bottom?
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 - top
\item \verb{[10]} 10 - bottom
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7514/vshow}{GSS Data Explorer page for \code{rank10fut}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab 1 - top \tab 10 - bottom \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2029 \tab 77 \tab 49 \tab 106 \tab 282 \tab 317 \tab 554 \tab 170 \tab 129 \tab 82 \tab 27 \tab 12 \tab 165 \tab 33 \tab 4032 \cr
   Total \tab 2029 \tab 77 \tab 49 \tab 106 \tab 282 \tab 317 \tab 554 \tab 170 \tab 129 \tab 82 \tab 27 \tab 12 \tab 165 \tab 33 \tab 4032 \cr
}
}

\keyword{variable}
