% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_631.R
\name{rank16}
\alias{rank16}
\title{R's self ranking of family at age 16}
\source{
General Social Survey https://gss.norc.org
}
\description{
rank16
}
\details{
Question In our society there are groups which tend to be towards the top and groups which tend to be towards the bottom. (CATI TEXT: On the handcard is ) (WEB TEXT: Below is ) a scale that runs from top to bottom.) If you think about the family that you grew up in, where did they fit in then?
}
\section{Values}{

\itemize{
\item \code{1} 1 - top
\item \code{10} 10 - bottom
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7513/vshow}{GSS Data Explorer page for \code{rank16}}.
}

\keyword{variable}
