% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_671.R
\name{ratedemp}
\alias{ratedemp}
\title{Rating of democratic party}
\source{
General Social Survey https://gss.norc.org
}
\description{
ratedemp
}
\details{
Question How would you rate the Democratic Party?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8097/vshow}{GSS Data Explorer page for \code{ratedemp}}.
}

\seealso{
Other ANES: 
\code{\link{lkelyvot}},
\code{\link{ratechall124}},
\code{\link{ratechall224}},
\code{\link{ratechall324}},
\code{\link{rateincumb24}},
\code{\link{ratemaga}},
\code{\link{raterepp}},
\code{\link{whovote24a}}
}
\concept{ANES}
\keyword{variable}
