% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_315.R
\name{ratepain}
\alias{ratepain}
\title{In the past 7 days, r's average pain}
\source{
General Social Survey https://gss.norc.org
}
\description{
ratepain
}
\details{
Question On a scale from 0 to 10, with 0 meaning no pain and 10 being the worst imaginable pain, how would you rate your pain on average?
}
\section{Values}{

\itemize{
\item \code{0} 0 - no pain
\item \code{10} 10 - the worst imaginable pain
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7171/vshow}{GSS Data Explorer page for \code{ratepain}}.
}

\keyword{variable}
