% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_434.R
\name{ratetone}
\alias{ratetone}
\title{R's facial coloring by interviewer}
\source{
General Social Survey https://gss.norc.org
}
\description{
ratetone
}
\details{
Question INTERVIEWER: PLEASE RECORD THE COLOR FROM THE COLOR CARD THAT MOST CLOSELY CORRESPONDS TO THE RESPONDENT’S FACIAL COLORING.
}
\section{Values}{

\itemize{
\item \code{1} 1 - lightest
\item \code{10} 10 - darkest
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3978/vshow}{GSS Data Explorer page for \code{ratetone}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 1 - lightest \tab 10 - darkest \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab no answer \tab not available in this year \tab Total \cr
   2012 \tab - \tab 561 \tab 18 \tab 531 \tab 234 \tab 106 \tab 72 \tab 45 \tab 65 \tab 49 \tab 17 \tab 276 \tab - \tab 1974 \cr
   2014 \tab - \tab 1045 \tab 6 \tab 513 \tab 247 \tab 102 \tab 86 \tab 60 \tab 58 \tab 39 \tab 28 \tab 354 \tab - \tab 2538 \cr
   2016 \tab - \tab 1049 \tab 9 \tab 785 \tab 292 \tab 166 \tab 116 \tab 82 \tab 79 \tab 61 \tab 21 \tab 207 \tab - \tab 2867 \cr
   2018 \tab - \tab 904 \tab 3 \tab 556 \tab 219 \tab 106 \tab 86 \tab 60 \tab 56 \tab 46 \tab 17 \tab 295 \tab - \tab 2348 \cr
   2022 \tab 1864 \tab 785 \tab 9 \tab 353 \tab 175 \tab 93 \tab 75 \tab 65 \tab 60 \tab 52 \tab 13 \tab - \tab - \tab 3544 \cr
   Total \tab 1864 \tab 4344 \tab 45 \tab 2738 \tab 1167 \tab 573 \tab 435 \tab 312 \tab 318 \tab 247 \tab 96 \tab 1132 \tab 0 \tab 13271 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{obeylaw}},
\code{\link{protest1}},
\code{\link{protest2}},
\code{\link{protest3}},
\code{\link{protest4}},
\code{\link{protest5}},
\code{\link{protest6}},
\code{\link{pubdef}},
\code{\link{pubecon}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
