% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_667.R
\name{rbioprnt}
\alias{rbioprnt}
\title{R is biological parent of child under 18 in hh}
\source{
General Social Survey https://gss.norc.org
}
\description{
rbioprnt
}
\details{
Question These next questions are about your relationships with other individuals who are under the age of 18 in this household.
Are you the biological parent of at least one individual under the age of 18 in this household?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8020/vshow}{GSS Data Explorer page for \code{rbioprnt}}.
}

\keyword{variable}
