% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_495.R
\name{rcontact}
\alias{rcontact}
\title{Do you have personal contact at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
rcontact
}
\details{
Question (HAND CARD D7)
(For each of these statements about your (main) job, please tell me how much you agree or disagree with it as it applies to your job.)
In my job, I have personal contact with other people.
(Would you say...)
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6329/vshow}{GSS Data Explorer page for \code{rcontact}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2016 \tab 1927 \tab 529 \tab 347 \tab 28 \tab 20 \tab 8 \tab 2 \tab 6 \tab 2867 \cr
   Total \tab 1927 \tab 529 \tab 347 \tab 28 \tab 20 \tab 8 \tab 2 \tab 6 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{chngewrk}},
\code{\link{decidwrk}},
\code{\link{harass5}},
\code{\link{nounemp5}},
\code{\link{nounemp6}},
\code{\link{unionsbd}},
\code{\link{wkathome}},
\code{\link{wkwkends}},
\code{\link{wrkshift}}
}
\concept{Work Orientation}
\keyword{variable}
