% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_495.R
\name{rcontact}
\alias{rcontact}
\title{Do you have personal contact at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
rcontact
}
\details{
Question (HAND CARD D7)
(For each of these statements about your (main) job, please tell me how much you agree or disagree with it as it applies to your job.)
In my job, I have personal contact with other people.
(Would you say...)
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6329/vshow}{GSS Data Explorer page for \code{rcontact}}.
}

\seealso{
Other Work Orientation: 
\code{\link{chngewrk}},
\code{\link{decidwrk}},
\code{\link{harass5}},
\code{\link{nounemp5}},
\code{\link{nounemp6}},
\code{\link{unionsbd}},
\code{\link{wkathome}},
\code{\link{wkwkends}},
\code{\link{wrkshift}}
}
\concept{Work Orientation}
\keyword{variable}
