% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_39.R
\name{rdeemlib}
\alias{rdeemlib}
\title{Contrast between redeemer and liberator}
\source{
General Social Survey https://gss.norc.org
}
\description{
rdeemlib
}
\details{
Question There are many different ways of picturing God. We'd like to know the kinds of images you are most likely to associate with God. Here is a card with sets of contrasting images. On a scale of 1-7 where would you place your image of God between the two contrasting images? (HAND CARD AE) The first set of contrasting images shows Mother at 1 on the scale and Father at 7. If you imagine God as a Mother you would place yourself at 1. If you imagine God as a Father, you would place yourself at 7. If you imagine God as somewhere between Mother and Father, you would place yourself at 2, 3, 4, 5, or 6. (REPEAT EXAMPLE AS NECESSARY FOR EACH ITEM A-F, SUBSTITUTING IMAGES A-F FOR "MOTHER" AND "FATHER".) Where would you place your image of God on the scale for . . . READ EACH SET OF IMAGES AND CIRCLE ONE CODE FOR EACH.
}
\section{Values}{

\itemize{
\item \code{1} redeemer
\item \code{7} liberator
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984, 1985, 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/339/vshow}{GSS Data Explorer page for \code{rdeemlib}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab redeemer \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab liberator \tab don't know \tab no answer \tab Total \cr
   1984 \tab 484 \tab 375 \tab 78 \tab 47 \tab 342 \tab 25 \tab 18 \tab 46 \tab 2 \tab 56 \tab 1473 \cr
   1985 \tab - \tab 524 \tab 109 \tab 101 \tab 553 \tab 41 \tab 33 \tab 94 \tab 21 \tab 58 \tab 1534 \cr
   1986 \tab - \tab 546 \tab 101 \tab 72 \tab 535 \tab 47 \tab 25 \tab 78 \tab 16 \tab 50 \tab 1470 \cr
   Total \tab 484 \tab 1445 \tab 288 \tab 220 \tab 1430 \tab 113 \tab 76 \tab 218 \tab 39 \tab 164 \tab 4477 \cr
}
}

\seealso{
Other Core: 
\code{\link{crtrheal}},
\code{\link{frndking}},
\code{\link{healer}},
\code{\link{judgeluv}},
\code{\link{king}},
\code{\link{liberatr}},
\code{\link{mapa}},
\code{\link{mastersp}},
\code{\link{popespks}}
}
\concept{Core}
\keyword{variable}
