% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_176.R
\name{rdiscaff}
\alias{rdiscaff}
\title{Chances family member not promoted for other race}
\source{
General Social Survey https://gss.norc.org
}
\description{
rdiscaff
}
\details{
Question A. What do you think the chances are these days that you or anyone in your family won't get a job or promotion while an equally or less qualified (White/(Black/African-American)) employee receives one instead?  Would you say the chances of this happening are very likely, somewhat likely, somewhat unlikely, or very unlikely?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} somewhat unlikely
\item \code{4} very unlikely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1647/vshow}{GSS Data Explorer page for \code{rdiscaff}}.
}

\seealso{
Other Multiculturalism: 
\code{\link{econfutr}},
\code{\link{futrdown}},
\code{\link{futrup}},
\code{\link{pastdown}},
\code{\link{pastup}},
\code{\link{rimmdisc}}

Other Job: 
\code{\link{rimmdisc}}
}
\concept{Affirmative Action}
\concept{Job}
\concept{Multiculturalism}
\keyword{variable}
