% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_194.R
\name{realtor1}
\alias{realtor1}
\title{What was the most important reason?}
\source{
General Social Survey https://gss.norc.org
}
\description{
realtor1
}
\details{
Question What made you choose the realtor with whom you worked? (Choose all that apply.)
G. MOST IMPORTANT REASON FROM ABOVE
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1824/vshow}{GSS Data Explorer page for \code{realtor1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab Total \cr
   1996 \tab 2864 \tab 17 \tab 7 \tab 1 \tab 5 \tab 5 \tab 5 \tab 2904 \cr
   Total \tab 2864 \tab 17 \tab 7 \tab 1 \tab 5 \tab 5 \tab 5 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{lawyer}},
\code{\link{lawyera}},
\code{\link{lawyerb}},
\code{\link{lawyerc}},
\code{\link{realtore}},
\code{\link{realtorf}},
\code{\link{relagent}},
\code{\link{sathome}},
\code{\link{warrntyh}}
}
\concept{Markets}
\keyword{variable}
