% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_524.R
\name{recycle}
\alias{recycle}
\title{Recycle cans bottles}
\source{
General Social Survey https://gss.norc.org
}
\description{
recycle
}
\details{
Question A. How often do you make a special effort to sort glass or cans or plastic or papers and so on for recycling?
}
\section{Values}{

\itemize{
\item \verb{[1]} always
\item \verb{[2]} often
\item \verb{[3]} sometimes
\item \verb{[4]} never
\item \verb{[5]} recycling not available where i live
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010, 2021 \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4769/vshow}{GSS Data Explorer page for \code{recycle}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab always \tab often \tab sometimes \tab never \tab recycling not available where i live \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1993 \tab 49 \tab 579 \tab 338 \tab 343 \tab 204 \tab 55 \tab 1 \tab 37 \tab - \tab 1606 \cr
   1994 \tab 1606 \tab 481 \tab 329 \tab 340 \tab 170 \tab 47 \tab - \tab 19 \tab - \tab 2992 \cr
   2000 \tab 1541 \tab 393 \tab 291 \tab 271 \tab 189 \tab 67 \tab 2 \tab 63 \tab - \tab 2817 \cr
   2010 \tab 614 \tab 563 \tab 324 \tab 317 \tab 190 \tab 28 \tab 1 \tab 7 \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 1006 \tab 278 \tab 218 \tab 198 \tab 130 \tab - \tab 192 \tab 7 \tab 4032 \cr
   Total \tab 5813 \tab 3022 \tab 1560 \tab 1489 \tab 951 \tab 327 \tab 4 \tab 318 \tab 7 \tab 13491 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{chemfree}},
\code{\link{comtype}},
\code{\link{drivless}},
\code{\link{grndemo}},
\code{\link{grngroup}},
\code{\link{grnmoney}},
\code{\link{grnsign}},
\code{\link{nomeat}},
\code{\link{usdoenuf}}
}
\concept{ISSP Environment}
\keyword{variable}
