% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_178.R
\name{refpromo}
\alias{refpromo}
\title{Refuse a job promotion?}
\source{
General Social Survey https://gss.norc.org
}
\description{
refpromo
}
\details{
Question In your (present/most recent) job, (have you ever done/did you ever do) any of the following because of your responsibilities to members of your family?
A. Refuse a job promotion?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1667/vshow}{GSS Data Explorer page for \code{refpromo}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 148 \tab 1192 \tab 1 \tab 119 \tab 2904 \cr
   Total \tab 1444 \tab 148 \tab 1192 \tab 1 \tab 119 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{balwkfam}},
\code{\link{earnmore}},
\code{\link{fairhwrk}},
\code{\link{famlife}},
\code{\link{noathome}},
\code{\link{refmorwk}},
\code{\link{refxhour}},
\code{\link{workless}},
\code{\link{worklife}}
}
\concept{Gender}
\keyword{variable}
