% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_533.R
\name{refugees}
\alias{refugees}
\title{Refugees should be allowed to stay in america?}
\source{
General Social Survey https://gss.norc.org
}
\description{
refugees
}
\details{
Question How much do you agree or disagree that refugees who have suffered political repression in their own country should be allowed to stay in America?
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} disagree strongly
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4864/vshow}{GSS Data Explorer page for \code{refugees}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 133 \tab 433 \tab 343 \tab 213 \tab 90 \tab 129 \tab 26 \tab 2904 \cr
   Total \tab 1537 \tab 133 \tab 433 \tab 343 \tab 213 \tab 90 \tab 129 \tab 26 \tab 2904 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{abroad}},
\code{\link{immameco}},
\code{\link{immideas}},
\code{\link{letin1}},
\code{\link{livecom1}},
\code{\link{meltpot1}},
\code{\link{res161}}
}
\concept{ISSP National Identity}
\keyword{variable}
