% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_533.R
\name{refugees}
\alias{refugees}
\title{Refugees should be allowed to stay in america?}
\source{
General Social Survey https://gss.norc.org
}
\description{
refugees
}
\details{
Question How much do you agree or disagree that refugees who have suffered political repression in their own country should be allowed to stay in America?
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} disagree strongly
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4864/vshow}{GSS Data Explorer page for \code{refugees}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{abroad}},
\code{\link{immameco}},
\code{\link{immideas}},
\code{\link{letin1}},
\code{\link{livecom1}},
\code{\link{meltpot1}},
\code{\link{res161}}

Other United States: 
\code{\link{immameco}},
\code{\link{immcrime}},
\code{\link{immideas}},
\code{\link{immjobs}},
\code{\link{letin1}},
\code{\link{meltpot1}}
}
\concept{ISSP National Identity}
\concept{United States}
\keyword{variable}
