% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_13.R
\name{reg16}
\alias{reg16}
\title{Region of residence, age 16}
\source{
General Social Survey https://gss.norc.org
}
\description{
reg16
}
\details{
Question In what state or foreign country were you living when you were 16 years old?
As of 2024, REG16 has been recoded to only account for four categories. This has been retroactively applied to all previous years of REG16. The original REG16 containing data from 1972 to 2022 has been renamed REG16_7222.
}
\section{Values}{

\itemize{
\item \verb{[0]} foreign
\item \verb{[1]} new england
\item \verb{[2]} middle atlantic
\item \verb{[3]} east north central
\item \verb{[4]} west north central
\item \verb{[5]} south atlantic
\item \verb{[6]} east south atlantic
\item \verb{[7]} west south central
\item \verb{[8]} mountain
\item \verb{[9]} pacific
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/84/vshow}{GSS Data Explorer page for \code{reg16}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab foreign \tab new england \tab middle atlantic \tab east north central \tab west north central \tab south atlantic \tab east south atlantic \tab west south central \tab mountain \tab pacific \tab iap \tab don't know \tab skipped on web \tab Total \cr
   1972 \tab 66 \tab 88 \tab 300 \tab 282 \tab 178 \tab 196 \tab 119 \tab 209 \tab 43 \tab 132 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 58 \tab 59 \tab 258 \tab 298 \tab 129 \tab 288 \tab 110 \tab 133 \tab 42 \tab 129 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 54 \tab 59 \tab 253 \tab 322 \tab 143 \tab 259 \tab 103 \tab 129 \tab 42 \tab 120 \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 50 \tab 55 \tab 260 \tab 311 \tab 125 \tab 272 \tab 120 \tab 127 \tab 57 \tab 113 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 66 \tab 64 \tab 263 \tab 315 \tab 143 \tab 250 \tab 111 \tab 122 \tab 51 \tab 114 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 73 \tab 69 \tab 243 \tab 357 \tab 112 \tab 279 \tab 100 \tab 120 \tab 47 \tab 130 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 61 \tab 67 \tab 269 \tab 327 \tab 137 \tab 272 \tab 107 \tab 129 \tab 42 \tab 121 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 64 \tab 53 \tab 237 \tab 304 \tab 129 \tab 243 \tab 127 \tab 126 \tab 59 \tab 126 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 73 \tab 56 \tab 297 \tab 384 \tab 137 \tab 372 \tab 186 \tab 168 \tab 56 \tab 131 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 76 \tab 92 \tab 271 \tab 344 \tab 161 \tab 217 \tab 125 \tab 126 \tab 63 \tab 124 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 62 \tab 85 \tab 222 \tab 309 \tab 142 \tab 208 \tab 136 \tab 126 \tab 63 \tab 120 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 66 \tab 100 \tab 215 \tab 270 \tab 162 \tab 215 \tab 148 \tab 148 \tab 75 \tab 135 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 72 \tab 85 \tab 222 \tab 306 \tab 145 \tab 195 \tab 133 \tab 117 \tab 73 \tab 122 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 67 \tab 76 \tab 264 \tab 356 \tab 167 \tab 336 \tab 217 \tab 135 \tab 73 \tab 128 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 76 \tab 83 \tab 220 \tab 288 \tab 155 \tab 220 \tab 110 \tab 132 \tab 74 \tab 123 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 82 \tab 75 \tab 237 \tab 312 \tab 147 \tab 230 \tab 116 \tab 118 \tab 69 \tab 151 \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 70 \tab 70 \tab 220 \tab 267 \tab 144 \tab 191 \tab 108 \tab 107 \tab 62 \tab 133 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 66 \tab 89 \tab 234 \tab 292 \tab 135 \tab 242 \tab 127 \tab 124 \tab 66 \tab 142 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 84 \tab 57 \tab 244 \tab 332 \tab 136 \tab 223 \tab 137 \tab 155 \tab 66 \tab 172 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 166 \tab 140 \tab 468 \tab 522 \tab 252 \tab 466 \tab 222 \tab 303 \tab 142 \tab 311 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 180 \tab 166 \tab 465 \tab 563 \tab 213 \tab 414 \tab 216 \tab 251 \tab 145 \tab 291 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 174 \tab 145 \tab 452 \tab 523 \tab 225 \tab 433 \tab 203 \tab 274 \tab 143 \tab 260 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 191 \tab 129 \tab 470 \tab 510 \tab 204 \tab 401 \tab 204 \tab 286 \tab 148 \tab 274 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 195 \tab 148 \tab 443 \tab 546 \tab 215 \tab 382 \tab 202 \tab 253 \tab 117 \tab 264 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 194 \tab 105 \tab 398 \tab 548 \tab 190 \tab 459 \tab 151 \tab 295 \tab 160 \tab 312 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 489 \tab 165 \tab 591 \tab 802 \tab 284 \tab 762 \tab 252 \tab 404 \tab 253 \tab 508 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 193 \tab 83 \tab 289 \tab 359 \tab 144 \tab 326 \tab 101 \tab 175 \tab 104 \tab 249 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 189 \tab 76 \tab 294 \tab 380 \tab 134 \tab 321 \tab 130 \tab 179 \tab 97 \tab 244 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 205 \tab 102 \tab 248 \tab 347 \tab 133 \tab 332 \tab 113 \tab 163 \tab 98 \tab 233 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 280 \tab 118 \tab 359 \tab 447 \tab 144 \tab 379 \tab 155 \tab 208 \tab 146 \tab 302 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 249 \tab 159 \tab 371 \tab 552 \tab 185 \tab 435 \tab 187 \tab 243 \tab 147 \tab 339 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 217 \tab 115 \tab 268 \tab 405 \tab 150 \tab 397 \tab 147 \tab 230 \tab 116 \tab 303 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab 176 \tab 543 \tab 812 \tab 335 \tab 536 \tab 235 \tab 356 \tab 242 \tab 783 \tab 14 \tab - \tab - \tab 4032 \cr
   2022 \tab 315 \tab 177 \tab 393 \tab 613 \tab 242 \tab 633 \tab 199 \tab 315 \tab 195 \tab 452 \tab - \tab 1 \tab 9 \tab 3544 \cr
   Total \tab 4523 \tab 3386 \tab 10781 \tab 13905 \tab 5777 \tab 11384 \tab 5157 \tab 6486 \tab 3376 \tab 7591 \tab 14 \tab 1 \tab 9 \tab 72390 \cr
}
}

\seealso{
Other Core: 
\code{\link{cobarate}},
\code{\link{coeftotlt}},
\code{\link{cogradtounder}},
\code{\link{mobile16}},
\code{\link{race}},
\code{\link{res16}},
\code{\link{sex}},
\code{\link{spbarate}},
\code{\link{spgradtounder}}
}
\concept{Core}
\keyword{variable}
