% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_301.R
\name{regdoc}
\alias{regdoc}
\title{R has regular doctor or clinic}
\source{
General Social Survey https://gss.norc.org
}
\description{
regdoc
}
\details{
Question This section is about physicians telling patients about how health insurance and managed care policies affect health care. First, I will ask you about your health care. Later, you will hear a recorded conversation between a doctor and patient and I'll ask some questions about the conversation.Do you have a regular doctor or clinic?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2738/vshow}{GSS Data Explorer page for \code{regdoc}}.
}

\seealso{
Other Doctors & Patients: 
\code{\link{docaskme}},
\code{\link{docdecid}},
\code{\link{docrely}},
\code{\link{docvig}},
\code{\link{regdocyr}}

Other Health: 
\code{\link{docaskme}},
\code{\link{docdecid}},
\code{\link{docrely}},
\code{\link{docvig}},
\code{\link{regdocyr}}
}
\concept{Doctors & Patients}
\concept{Health}
\keyword{variable}
