% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_301.R
\name{regdocyr}
\alias{regdocyr}
\title{R has had doctor for more than one year}
\source{
General Social Survey https://gss.norc.org
}
\description{
regdocyr
}
\details{
Question Have you been a patient with this doctor or clinic for more than one year?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2739/vshow}{GSS Data Explorer page for \code{regdocyr}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2002 \tab 467 \tab 2025 \tab 269 \tab 4 \tab 2765 \cr
   Total \tab 467 \tab 2025 \tab 269 \tab 4 \tab 2765 \cr
}
}

\seealso{
Other Doctors & Patients: 
\code{\link{docaskme}},
\code{\link{docdecid}},
\code{\link{docrely}},
\code{\link{docvig}},
\code{\link{regdoc}}
}
\concept{Doctors & Patients}
\keyword{variable}
