% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_513.R
\name{relactiv}
\alias{relactiv}
\title{How often does r take part in relig activities}
\source{
General Social Survey https://gss.norc.org
}
\description{
relactiv
}
\details{
Question How often do you take part in the activities and organizations of a church or place of worship other than attending services?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} less than once a year
\item \code{3} about once or twice a year
\item \code{4} several times a year
\item \code{5} about once a month
\item \code{6} 2-3 times a month
\item \code{7} nearly every week
\item \code{8} every week
\item \code{9} several times a week
\item \code{10} once a day
\item \code{11} several times a day
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4677/vshow}{GSS Data Explorer page for \code{relactiv}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{densp}},
\code{\link{feelrel}},
\code{\link{prayfreq}},
\code{\link{relactiv1}},
\code{\link{schlpray}}

Other Religion: 
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{densp}},
\code{\link{feelrel}},
\code{\link{prayfreq}},
\code{\link{relactiv1}},
\code{\link{schlpray}}
}
\concept{ISSP Religion}
\concept{Religion}
\keyword{variable}
