% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_194.R
\name{relagent}
\alias{relagent}
\title{Your relationship to the real estate agent}
\source{
General Social Survey https://gss.norc.org
}
\description{
relagent
}
\details{
Question Which of the following best describes your relationship to the real estate agent with whom you worked in your purchase of the home or to the owner of the real estate agency? Please indicate what your closest relationship was.
}
\section{Values}{

\itemize{
\item \code{1} a relative including in-laws
\item \code{2} a friend or acquaintance
\item \code{3} a friend of a friend-relative
\item \code{4} had previous business dealings
\item \code{5} no prior relationship
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1825/vshow}{GSS Data Explorer page for \code{relagent}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab a relative including in-laws \tab a friend or acquaintance \tab a friend of a friend-relative \tab had previous business dealings \tab no prior relationship \tab Total \cr
   1996 \tab 2686 \tab 10 \tab 35 \tab 15 \tab 14 \tab 144 \tab 2904 \cr
   Total \tab 2686 \tab 10 \tab 35 \tab 15 \tab 14 \tab 144 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{lawyer}},
\code{\link{lawyera}},
\code{\link{lawyerb}},
\code{\link{lawyerc}},
\code{\link{realtor1}},
\code{\link{realtore}},
\code{\link{realtorf}},
\code{\link{sathome}},
\code{\link{warrntyh}}
}
\concept{Markets}
\keyword{variable}
