% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_343.R
\name{relalt1}
\alias{relalt1}
\title{Rel change altered r's outlook on life}
\source{
General Social Survey https://gss.norc.org
}
\description{
relalt1
}
\details{
Question To what extent did the religious/spiritual change you just told me about do the following?
A. Altered your outlook on life?
}
\section{Values}{

\itemize{
\item \code{1} not at all
\item \code{7} a great deal
\item \code{9} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3077/vshow}{GSS Data Explorer page for \code{relalt1}}.
}

\seealso{
Other Spirituality and Religion: 
\code{\link{agerborn}},
\code{\link{agerelex}},
\code{\link{alloflfe}},
\code{\link{blessngs}},
\code{\link{changed1}},
\code{\link{changed2}},
\code{\link{changed3}},
\code{\link{chngrbrn}},
\code{\link{chngrel}},
\code{\link{closrgod}},
\code{\link{entity}},
\code{\link{expchng1}},
\code{\link{expchng2}},
\code{\link{expchng3}},
\code{\link{godclose}},
\code{\link{godguide}},
\code{\link{godhelp}},
\code{\link{godlvdir}},
\code{\link{godlvoth}},
\code{\link{idols}},
\code{\link{joylifts}},
\code{\link{notthink}},
\code{\link{numrborn}},
\code{\link{numrelex}},
\code{\link{relalt2}},
\code{\link{relalt3}},
\code{\link{relalt4}},
\code{\link{relalt5}},
\code{\link{relalt6}},
\code{\link{relsprt1}},
\code{\link{relsprt2}},
\code{\link{rosaries}},
\code{\link{whychng1}},
\code{\link{whychng2}},
\code{\link{whychng3}}
}
\concept{Spirituality and Religion}
\keyword{variable}
