% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_232.R
\name{relart}
\alias{relart}
\title{Had experience of art at place of worship}
\source{
General Social Survey https://gss.norc.org
}
\description{
relart
}
\details{
Question Within the past twelve months, have you had any experiences with the arts at your place of worship?  I mean activities like viewing performances or exhibits, singing in a choir, or playing an instrument.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2132/vshow}{GSS Data Explorer page for \code{relart}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1998 \tab 1797 \tab 313 \tab 643 \tab 79 \tab 2832 \cr
   Total \tab 1797 \tab 313 \tab 643 \tab 79 \tab 2832 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{aimofart}},
\code{\link{drama}},
\code{\link{localart}},
\code{\link{natlart}},
\code{\link{popmusic}},
\code{\link{readfict}},
\code{\link{stateart}},
\code{\link{volarts}}
}
\concept{Religion and Culture}
\keyword{variable}
