% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_232.R
\name{relart}
\alias{relart}
\title{Had experience of art at place of worship}
\source{
General Social Survey https://gss.norc.org
}
\description{
relart
}
\details{
Question Within the past twelve months, have you had any experiences with the arts at your place of worship?  I mean activities like viewing performances or exhibits, singing in a choir, or playing an instrument.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2132/vshow}{GSS Data Explorer page for \code{relart}}.
}

\seealso{
Other Religion and Culture: 
\code{\link{aimofart}},
\code{\link{drama}},
\code{\link{localart}},
\code{\link{natlart}},
\code{\link{popmusic}},
\code{\link{readfict}},
\code{\link{stateart}},
\code{\link{volarts}}

Other Leisure: 
\code{\link{drama}},
\code{\link{popmusic}},
\code{\link{readfict}},
\code{\link{volarts}}
}
\concept{Leisure}
\concept{Religion and Culture}
\keyword{variable}
