% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_583.R
\name{relate3}
\alias{relate3}
\title{Relationship of 3rd person to household head}
\source{
General Social Survey https://gss.norc.org
}
\description{
relate3
}
\details{
Question Please tell me the names of the people who usually live in this household. Let's start with the head of the household. (LIST ON LINES 01-10 BELOW.)
Have we forgotten anyone: such as babies or small children; roomers; people who usually live here, but are away temporarily--on business trips, vacations, at school, temporarily in a hospital, and so on? (LIST ADDITIONAL PERSONS ON LINES 01-10 BELOW.)
Are there any people currently staying here--visitors, friends or relatives--do not usually live here? (LIST VISITORS ON LINES 11-14 BELOW.)
A. What is that (PERSON'S) relationship to the HEAD OF HOUSEHOLD?
B. Code sex. (ASK IF NOT OBVIOUS).
C. How old was (HEAD/PERSON) on (his/her) last birthday?
D. IF 13 YRS OR OLDER ASK: Is (PERSON) now married, widowed, divorced, separated, or has (he/she) never been married?
E. Are any of the people we have listed staying somewhere else right now? If yes, who is staying somewhere else right now?
F. Where is (PERSON) staying right now? Is (PERSON) staying at another household;
Is (he/she) traveling; Is (he/she) in some institution or dormitory--like at college or in a hospital or something; or what?
\enumerate{
\item Relationship of third person to head of household
}
}
\section{Values}{

\itemize{
\item \code{1} head of household
\item \code{2} spouse
\item \code{3} child (natural, adopted, step)
\item \code{4} son-/daughter-in-law
\item \code{5} grand/great-grandchild
\item \code{6} parent/parent-in-law
\item \code{7} other relative
\item \code{8} non-relative
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5344/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab child (natural, adopted, step) \tab grand/great-grandchild \tab non-relative \tab other relative \tab parent/parent-in-law \tab son-/daughter-in-law \tab spouse \tab no answer \tab head of household \tab skipped on web \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 656 \tab 768 \tab 7 \tab 22 \tab 17 \tab 11 \tab 6 \tab 3 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1134 \tab 333 \tab 7 \tab 8 \tab 9 \tab 5 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 724 \tab 742 \tab 11 \tab 9 \tab 16 \tab 18 \tab 2 \tab 3 \tab 5 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 749 \tab 714 \tab 9 \tab 17 \tab 18 \tab 11 \tab 4 \tab 3 \tab 7 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 780 \tab 597 \tab 10 \tab 24 \tab 14 \tab 11 \tab 8 \tab - \tab 24 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 991 \tab 766 \tab 27 \tab 20 \tab 16 \tab 18 \tab 5 \tab 5 \tab 12 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 797 \tab 723 \tab 14 \tab 17 \tab 21 \tab 10 \tab 5 \tab 7 \tab 3 \tab 2 \tab - \tab - \tab 1599 \cr
   1984 \tab 783 \tab 619 \tab 10 \tab 25 \tab 11 \tab 11 \tab 1 \tab 5 \tab 8 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 844 \tab 625 \tab 5 \tab 22 \tab 10 \tab 13 \tab 2 \tab 9 \tab 4 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 768 \tab 612 \tab 12 \tab 29 \tab 15 \tab 14 \tab 6 \tab 8 \tab 6 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 948 \tab 707 \tab 27 \tab 47 \tab 28 \tab 11 \tab 2 \tab 10 \tab 37 \tab 2 \tab - \tab - \tab 1819 \cr
   1988 \tab 847 \tab 548 \tab 14 \tab 22 \tab 27 \tab 9 \tab 2 \tab 3 \tab 9 \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 834 \tab 592 \tab 12 \tab 29 \tab 11 \tab 4 \tab 6 \tab 8 \tab 40 \tab 1 \tab - \tab - \tab 1537 \cr
   1990 \tab 810 \tab 491 \tab 10 \tab 23 \tab 17 \tab 6 \tab 5 \tab 4 \tab 6 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 849 \tab 577 \tab 10 \tab 34 \tab 22 \tab 10 \tab 1 \tab 5 \tab 9 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 895 \tab 630 \tab 9 \tab 28 \tab 17 \tab 11 \tab 7 \tab 5 \tab 4 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1751 \tab 1094 \tab 23 \tab 47 \tab 32 \tab 21 \tab 4 \tab 13 \tab 7 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1729 \tab 970 \tab 21 \tab 63 \tab 24 \tab 20 \tab 10 \tab 13 \tab 54 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 1713 \tab 923 \tab 15 \tab 56 \tab 32 \tab 12 \tab 10 \tab 23 \tab 48 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 1668 \tab 924 \tab 33 \tab 56 \tab 37 \tab 29 \tab 10 \tab - \tab 60 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1815 \tab 774 \tab 39 \tab 63 \tab 40 \tab 13 \tab 15 \tab - \tab 6 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1780 \tab 899 \tab 34 \tab 31 \tab 29 \tab 24 \tab 11 \tab - \tab 4 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2778 \tab 1495 \tab 44 \tab 84 \tab 52 \tab 45 \tab 9 \tab - \tab 3 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1221 \tab 684 \tab 34 \tab 29 \tab 33 \tab 18 \tab 1 \tab - \tab 3 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1319 \tab 600 \tab 26 \tab 45 \tab 26 \tab 15 \tab 10 \tab - \tab 3 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1259 \tab 583 \tab 14 \tab 41 \tab 47 \tab 20 \tab 3 \tab - \tab 7 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 1642 \tab 743 \tab 21 \tab 68 \tab 29 \tab 22 \tab 8 \tab - \tab 5 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 1919 \tab 765 \tab 37 \tab 59 \tab 37 \tab 33 \tab 14 \tab - \tab 3 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 1561 \tab 651 \tab 28 \tab 46 \tab 31 \tab 23 \tab 7 \tab - \tab 1 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 2440 \tab 1214 \tab 52 \tab 150 \tab 84 \tab 72 \tab 3 \tab - \tab - \tab - \tab 17 \tab - \tab 4032 \cr
   2022 \tab 2985 \tab 447 \tab 11 \tab 32 \tab 31 \tab 27 \tab 7 \tab - \tab 3 \tab - \tab 1 \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 45590 \tab 22810 \tab 626 \tab 1246 \tab 833 \tab 567 \tab 185 \tab 129 \tab 381 \tab 5 \tab 18 \tab 3309 \tab 75699 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away3}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{mar3}},
\code{\link{old3}},
\code{\link{old4}},
\code{\link{relate4}},
\code{\link{where2}},
\code{\link{where3}}

Other Cohabitation: 
\code{\link{relate4}}

Other Family: 
\code{\link{away3}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{mar3}},
\code{\link{old3}},
\code{\link{old4}},
\code{\link{relate4}},
\code{\link{where2}},
\code{\link{where3}}

Other Household: 
\code{\link{away3}},
\code{\link{gender3}},
\code{\link{gender4}},
\code{\link{mar3}},
\code{\link{old3}},
\code{\link{old4}},
\code{\link{relate4}},
\code{\link{where2}},
\code{\link{where3}}
}
\concept{Cohabitation}
\concept{Family}
\concept{Household}
\concept{Household Members and Structure}
\keyword{variable}
