% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_554.R
\name{relatsex}
\alias{relatsex}
\title{Relation to last sex partner}
\source{
General Social Survey https://gss.norc.org
}
\description{
relatsex
}
\details{
Question The last time you had sex, was it with someone you were in an on-going relationship with, or was it with someone else? Remember that by "sex" we mean only vaginal, oral, or anal sex.
}
\section{Values}{

\itemize{
\item \code{1} yes, the last time i had sex, it was with someone i was in an on-going relationship with
\item \code{2} no, the last time i had sex, it was not with someone i was in an on-going relationship with
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab B/C/- \tab partial \cr
   2018 \tab B/C/- \tab partial \cr
   2021 \tab B/C/- \tab partial \cr
   2022 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5069/vshow}{GSS Data Explorer page for \code{relatsex}}.
}

\seealso{
Other High Risk Behaviors: 
\code{\link{aidslook}},
\code{\link{condom}},
\code{\link{crack30}},
\code{\link{evcrack}},
\code{\link{evidu}},
\code{\link{hivtest}},
\code{\link{hivtest1}},
\code{\link{hivtest2}},
\code{\link{idu30}}
}
\concept{High Risk Behaviors}
\keyword{variable}
