% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_454.R
\name{relcall}
\alias{relcall}
\title{Contact with most contacted rel}
\source{
General Social Survey https://gss.norc.org
}
\description{
relcall
}
\details{
Question Thinking of all the adult relatives, which one do you have the most contact with?
C. And how often do you have any other contact with your relative besides visiting, either by telephone or letter?
}
\section{Values}{

\itemize{
\item \code{1} daily
\item \code{2} several x week
\item \code{3} once a week
\item \code{4} once a month
\item \code{5} several x year
\item \code{6} less often
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1986 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4161/vshow}{GSS Data Explorer page for \code{relcall}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{chores1}},
\code{\link{chores2}},
\code{\link{fricall}},
\code{\link{frineigh}},
\code{\link{frinum}},
\code{\link{frisex}},
\code{\link{fritime}},
\code{\link{frivisit}},
\code{\link{friwork}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
