% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_516.R
\name{relcand}
\alias{relcand}
\title{R accepts a person from a different religion or view from r's being candidate of}
\source{
General Social Survey https://gss.norc.org
}
\description{
relcand
}
\details{
Question People belong to different religions and have different religious views.
B. Would you accept a person from a different religion or with a very different religious view from yours being a candidate of the political party you prefer?
}
\section{Values}{

\itemize{
\item \code{1} definitely accept
\item \code{2} probably accept
\item \code{3} probably not accept
\item \code{4} definitely not accept
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4702/vshow}{GSS Data Explorer page for \code{relcand}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{ancestrs}},
\code{\link{mywaygod}},
\code{\link{nirvana}},
\code{\link{reincar}},
\code{\link{relext1}},
\code{\link{relext2}},
\code{\link{relmarry}}
}
\concept{ISSP Religion}
\keyword{variable}
