% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_228.R
\name{relcmfrt}
\alias{relcmfrt}
\title{Find strength and comfort in religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
relcmfrt
}
\details{
Question The following questions deal with possible daily spiritual experiences. To what extent can you say you experience the following:
B. I find strength and comfort in my religion.
}
\section{Values}{

\itemize{
\item \code{1} many times a day
\item \code{2} every day
\item \code{3} most days
\item \code{4} some days
\item \code{5} once in a while
\item \code{6} never/almst nevr
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2109/vshow}{GSS Data Explorer page for \code{relcmfrt}}.
}

\seealso{
Other Religion and Culture: 
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{feelgod}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{godwatch}},
\code{\link{lesspain}},
\code{\link{rellife}}
}
\concept{Religion and Culture}
\keyword{variable}
