% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_192.R
\name{reldealr}
\alias{reldealr}
\title{R relationship to the salesperson}
\source{
General Social Survey https://gss.norc.org
}
\description{
reldealr
}
\details{
Question Which of the following best describes your relationship to the salesperson from whom you purchased the car or to the owner of the auto dealership. Please indicate what your closest relationship was.
}
\section{Values}{

\itemize{
\item \verb{[1]} a relative , including in-laws
\item \verb{[2]} a friend or acquaintance
\item \verb{[3]} a friend of a friend-relative
\item \verb{[4]} previous business dealings
\item \verb{[5]} no prior relationship
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1810/vshow}{GSS Data Explorer page for \code{reldealr}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a relative , including in-laws \tab a friend or acquaintance \tab a friend of a friend-relative \tab previous business dealings \tab no prior relationship \tab no answer \tab Total \cr
   1996 \tab 2497 \tab 9 \tab 45 \tab 22 \tab 41 \tab 279 \tab 11 \tab 2904 \cr
   Total \tab 2497 \tab 9 \tab 45 \tab 22 \tab 41 \tab 279 \tab 11 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{carbuy1}},
\code{\link{carbuya}},
\code{\link{carbuyb}},
\code{\link{carbuyc}},
\code{\link{carbuyd}},
\code{\link{carbuye}},
\code{\link{carbuyf}},
\code{\link{carbuyg}},
\code{\link{satcar}}
}
\concept{Markets}
\keyword{variable}
