% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_192.R
\name{reldealr}
\alias{reldealr}
\title{R relationship to the salesperson}
\source{
General Social Survey https://gss.norc.org
}
\description{
reldealr
}
\details{
Question Which of the following best describes your relationship to the salesperson from whom you purchased the car or to the owner of the auto dealership. Please indicate what your closest relationship was.
}
\section{Values}{

\itemize{
\item \code{1} a relative , including in-laws
\item \code{2} a friend or acquaintance
\item \code{3} a friend of a friend-relative
\item \code{4} previous business dealings
\item \code{5} no prior relationship
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1810/vshow}{GSS Data Explorer page for \code{reldealr}}.
}

\seealso{
Other Markets: 
\code{\link{carbuy1}},
\code{\link{carbuya}},
\code{\link{carbuyb}},
\code{\link{carbuyc}},
\code{\link{carbuyd}},
\code{\link{carbuye}},
\code{\link{carbuyf}},
\code{\link{carbuyg}},
\code{\link{satcar}}

Other Purchases/Transactions: 
\code{\link{carbuy1}},
\code{\link{carbuya}},
\code{\link{carbuyb}},
\code{\link{carbuyc}},
\code{\link{carbuyd}},
\code{\link{carbuye}},
\code{\link{carbuyf}},
\code{\link{carbuyg}},
\code{\link{satcar}}
}
\concept{Markets}
\concept{Purchases/Transactions}
\keyword{variable}
