% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_231.R
\name{relexp}
\alias{relexp}
\title{Have religious experience changed life}
\source{
General Social Survey https://gss.norc.org
}
\description{
relexp
}
\details{
Question Did you ever have a religious or spiritual experience that changed your life?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2021 \cr
   A/B/C \tab 1998, 2004, 2006, 2008, 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2122/vshow}{GSS Data Explorer page for \code{relexp}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1998 \tab 1387 \tab 557 \tab 867 \tab 5 \tab 16 \tab - \tab 2832 \cr
   2004 \tab 1472 \tab 477 \tab 848 \tab 4 \tab 11 \tab - \tab 2812 \cr
   2006 \tab 1518 \tab 1237 \tab 1730 \tab 18 \tab 7 \tab - \tab 4510 \cr
   2008 \tab - \tab 869 \tab 1135 \tab 12 \tab 7 \tab - \tab 2023 \cr
   2010 \tab - \tab 836 \tab 1188 \tab 11 \tab 9 \tab - \tab 2044 \cr
   2021 \tab 2687 \tab 541 \tab 789 \tab 2 \tab 1 \tab 12 \tab 4032 \cr
   Total \tab 7064 \tab 4517 \tab 6557 \tab 52 \tab 51 \tab 12 \tab 18253 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{artgod}},
\code{\link{artists}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{natarts}},
\code{\link{relpersn}},
\code{\link{sprtprsn}}
}
\concept{Religion and Culture}
\keyword{variable}
