% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_512.R
\name{relexper}
\alias{relexper}
\title{Turning point in life for religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
relexper
}
\details{
Question Has there ever been a turning point in your life when you made a new and personal commitment to religion?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991, 1998, 2004, 2006, 2008, 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4664/vshow}{GSS Data Explorer page for \code{relexper}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 615 \tab 698 \tab 3 \tab 43 \tab 1517 \cr
   1998 \tab 1548 \tab 562 \tab 656 \tab 5 \tab 61 \tab 2832 \cr
   2004 \tab 1472 \tab 488 \tab 840 \tab 4 \tab 8 \tab 2812 \cr
   2006 \tab 1518 \tab 1312 \tab 1664 \tab 10 \tab 6 \tab 4510 \cr
   2008 \tab - \tab 860 \tab 1141 \tab 16 \tab 6 \tab 2023 \cr
   2010 \tab - \tab 847 \tab 1184 \tab 4 \tab 9 \tab 2044 \cr
   Total \tab 4696 \tab 4684 \tab 6183 \tab 42 \tab 133 \tab 15738 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{denkid}},
\code{\link{egomeans}},
\code{\link{madenkid}},
\code{\link{marelkid}},
\code{\link{ownfate}},
\code{\link{padenkid}},
\code{\link{parelkid}},
\code{\link{religkid}},
\code{\link{religsp}}
}
\concept{ISSP Religion}
\keyword{variable}
