% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_196.R
\name{relfixer}
\alias{relfixer}
\title{Relationship to the person who did the work}
\source{
General Social Survey https://gss.norc.org
}
\description{
relfixer
}
\details{
Question What was your relationship to the person who did the work, or to his or her boss (whichever relationship is closer)?
}
\section{Values}{

\itemize{
\item \code{1} a relative including in-laws
\item \code{2} a friend or acquaintance
\item \code{3} a friend of a friend-relative, etc
\item \code{4} previous business dealings
\item \code{5} no prior relationship
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1847/vshow}{GSS Data Explorer page for \code{relfixer}}.
}

\seealso{
Other Markets: 
\code{\link{borrowed}},
\code{\link{fixer1}},
\code{\link{fixerc}},
\code{\link{fixerd}},
\code{\link{fixere}},
\code{\link{fixerf}},
\code{\link{satfixes}},
\code{\link{sellbed}},
\code{\link{wholoand}}
}
\concept{Markets}
\keyword{variable}
