% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_196.R
\name{relfixer}
\alias{relfixer}
\title{Relationship to the person who did the work}
\source{
General Social Survey https://gss.norc.org
}
\description{
relfixer
}
\details{
Question What was your relationship to the person who did the work, or to his or her boss (whichever relationship is closer)?
}
\section{Values}{

\itemize{
\item \verb{[1]} a relative including in-laws
\item \verb{[2]} a friend or acquaintance
\item \verb{[3]} a friend of a friend-relative, etc
\item \verb{[4]} previous business dealings
\item \verb{[5]} no prior relationship
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1847/vshow}{GSS Data Explorer page for \code{relfixer}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a relative including in-laws \tab a friend or acquaintance \tab a friend of a friend-relative, etc \tab previous business dealings \tab no prior relationship \tab no answer \tab Total \cr
   1996 \tab 2353 \tab 18 \tab 92 \tab 39 \tab 65 \tab 333 \tab 4 \tab 2904 \cr
   Total \tab 2353 \tab 18 \tab 92 \tab 39 \tab 65 \tab 333 \tab 4 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{borrowed}},
\code{\link{fixer1}},
\code{\link{fixerc}},
\code{\link{fixerd}},
\code{\link{fixere}},
\code{\link{fixerf}},
\code{\link{satfixes}},
\code{\link{sellbed}},
\code{\link{wholoand}}
}
\concept{Markets}
\keyword{variable}
