% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_517.R
\name{relgenbar}
\alias{relgenbar}
\title{Religions are usually a barrier to equality between women and men}
\source{
General Social Survey https://gss.norc.org
}
\description{
relgenbar
}
\details{
Question Religions are usually a barrier to equality between women and men.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7281/vshow}{GSS Data Explorer page for \code{relgenbar}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2018 \tab 1173 \tab 62 \tab 255 \tab 305 \tab 415 \tab 101 \tab 23 \tab 14 \tab 2348 \cr
   Total \tab 1173 \tab 62 \tab 255 \tab 305 \tab 415 \tab 101 \tab 23 \tab 14 \tab 2348 \cr
}
}

\keyword{variable}
