% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_427.R
\name{relgendr}
\alias{relgendr}
\title{What is your relationship situation}
\source{
General Social Survey https://gss.norc.org
}
\description{
relgendr
}
\details{
Question Which of the following best describes your current relationship situation?
}
\section{Values}{

\itemize{
\item \verb{[1]} i am in a steady relationship with a woman and we live together
\item \verb{[2]} i am in a steady relationship with a man and we live together
\item \verb{[3]} i am in a steady relationship with a woman and we do not live together
\item \verb{[4]} i am in a steady relationship with a man and we do not live together
\item \verb{[5]} i am not in a steady relationship right now
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3910/vshow}{GSS Data Explorer page for \code{relgendr}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab i am in a steady relationship with a woman and we live together \tab i am in a steady relationship with a man and we live together \tab i am in a steady relationship with a woman and we do not live together \tab i am in a steady relationship with a man and we do not live together \tab i am not in a steady relationship right now \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1879 \tab 25 \tab 24 \tab 5 \tab 9 \tab 38 \tab 5 \tab 38 \tab 2023 \cr
   Total \tab 1879 \tab 25 \tab 24 \tab 5 \tab 9 \tab 38 \tab 5 \tab 38 \tab 2023 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{adoptkid}},
\code{\link{attractd}},
\code{\link{evkid}},
\code{\link{kidlived}},
\code{\link{kidnow1}},
\code{\link{kidresp}},
\code{\link{marunion}},
\code{\link{sexunion}},
\code{\link{toldsxor}}
}
\concept{Sexual Orientation}
\keyword{variable}
