% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_516.R
\name{relgeneq}
\alias{relgeneq}
\title{R's religion treats men and women equally}
\source{
General Social Survey https://gss.norc.org
}
\description{
relgeneq
}
\details{
Question In your opinion, does your religion treat men and women equally, treat men better than women, or treat women better than men?
}
\section{Values}{

\itemize{
\item \verb{[1]} i don't belong to or follow any religion
\item \verb{[2]} treats men and women equally
\item \verb{[3]} treats men better than women
\item \verb{[4]} treats women better than men
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7280/vshow}{GSS Data Explorer page for \code{relgeneq}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab i don't belong to or follow any religion \tab treats men and women equally \tab treats men better than women \tab treats women better than men \tab can't choose \tab no answer \tab Total \cr
   2018 \tab 1173 \tab 167 \tab 773 \tab 161 \tab 15 \tab 48 \tab 11 \tab 2348 \cr
   Total \tab 1173 \tab 167 \tab 773 \tab 161 \tab 15 \tab 48 \tab 11 \tab 2348 \cr
}
}

\keyword{variable}
