% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_44.R
\name{relgrade}
\alias{relgrade}
\title{Grade given religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
relgrade
}
\details{
Question Using the school grading system of A,B,C,D, and F for Fail, how Do you rate your church or religious group in meeting your spiritual needs?
}
\section{Values}{

\itemize{
\item \verb{[1]} a
\item \verb{[2]} b
\item \verb{[3]} c
\item \verb{[4]} d
\item \verb{[5]} f
\item \verb{[6]} no religion
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/381/vshow}{GSS Data Explorer page for \code{relgrade}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab a \tab b \tab c \tab d \tab f \tab no religion \tab don't know \tab no answer \tab Total \cr
   1989 \tab 531 \tab 278 \tab 287 \tab 182 \tab 45 \tab 33 \tab 125 \tab 51 \tab 5 \tab 1537 \cr
   Total \tab 531 \tab 278 \tab 287 \tab 182 \tab 45 \tab 33 \tab 125 \tab 51 \tab 5 \tab 1537 \cr
}
}

\seealso{
Other Core: 
\code{\link{cathtemp}},
\code{\link{feclergy}},
\code{\link{fepriest}},
\code{\link{jewtemp}},
\code{\link{mslmtemp}},
\code{\link{racdin}},
\code{\link{racmar}},
\code{\link{racmar10}},
\code{\link{racpush}}
}
\concept{Core}
\keyword{variable}
