% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_593.R
\name{relhh5}
\alias{relhh5}
\title{Relationship of person 5 to head of household}
\source{
General Social Survey https://gss.norc.org
}
\description{
relhh5
}
\details{
Question What is (PERSON)’s relationship to head of household?
E. Relationship of fifth person to head of household?
}
\section{Values}{

\itemize{
\item \code{1} head of household
\item \code{2} spouse
\item \code{3} partner, fiance(e), boyfriend, girlfriend, etc.
\item \code{4} child, not specified
\item \code{5} grandchild, great-grandchild
\item \code{6} son-in-law, daughter-in-law
\item \code{7} parent
\item \code{8} parent-in-law
\item \code{9} grandparent, great-grandparent
\item \code{10} uncle or aunt
\item \code{11} niece or nephew
\item \code{12} cousin
\item \code{13} brother, sister (full, half, or step sibling)
\item \code{14} sibling-in-law
\item \code{15} other relative (great-aunt, grandniece, etc.)
\item \code{16} roommate, housemate
\item \code{17} friend
\item \code{18} child of non-relative
\item \code{19} other non-relative (employee, boarder)
\item \code{41} natural/biological child
\item \code{42} adopted child
\item \code{43} stepchild
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5443/vshow}{GSS Data Explorer page for \code{relhh5}}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab iap \tab child, not specified \tab grandchild, great-grandchild \tab parent \tab uncle or aunt \tab niece or nephew \tab cousin \tab brother, sister (full, half, or step sibling) \tab sibling-in-law \tab roommate, housemate \tab friend \tab child of non-relative \tab other non-relative (employee, boarder) \tab natural/biological child \tab adopted child \tab stepchild \tab no answer \tab grandparent, great-grandparent \tab other relative (great-aunt, grandniece, etc.) \tab son-in-law, daughter-in-law \tab parent-in-law \tab partner, fiance(e), boyfriend, girlfriend, etc. \tab don't know \tab not available in this release \tab Total \cr
   2000 \tab 2559 \tab 25 \tab 20 \tab 3 \tab 1 \tab 7 \tab 1 \tab 2 \tab 2 \tab 3 \tab 2 \tab 4 \tab 1 \tab 158 \tab 7 \tab 7 \tab 15 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2556 \tab 37 \tab 18 \tab - \tab - \tab 3 \tab - \tab 3 \tab - \tab 2 \tab 1 \tab 2 \tab 3 \tab 129 \tab 1 \tab 5 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2567 \tab 61 \tab 24 \tab 3 \tab 1 \tab 5 \tab - \tab - \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 127 \tab - \tab 10 \tab - \tab - \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4086 \tab 131 \tab 26 \tab 2 \tab - \tab 5 \tab - \tab 6 \tab - \tab 5 \tab 2 \tab 1 \tab 4 \tab 218 \tab 3 \tab 13 \tab - \tab - \tab 1 \tab 3 \tab - \tab 3 \tab 1 \tab - \tab 4510 \cr
   2008 \tab 1826 \tab 84 \tab 22 \tab - \tab - \tab 2 \tab - \tab 1 \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab 69 \tab - \tab 9 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab 2023 \cr
   2010 \tab 1861 \tab 80 \tab 25 \tab 1 \tab 1 \tab 5 \tab - \tab 1 \tab 2 \tab 5 \tab 2 \tab - \tab - \tab 46 \tab 3 \tab 4 \tab 3 \tab - \tab 3 \tab 2 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1775 \tab 90 \tab 26 \tab 2 \tab 2 \tab 1 \tab - \tab 2 \tab 2 \tab 4 \tab - \tab 2 \tab 1 \tab 56 \tab - \tab 8 \tab - \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab 1974 \cr
   2014 \tab 2325 \tab 103 \tab 18 \tab 1 \tab 1 \tab 3 \tab 2 \tab 2 \tab - \tab 4 \tab 1 \tab 1 \tab - \tab 66 \tab 2 \tab 5 \tab - \tab - \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab 2538 \cr
   2016 \tab 2632 \tab 132 \tab 30 \tab 2 \tab - \tab 6 \tab - \tab 3 \tab 3 \tab 5 \tab 1 \tab 3 \tab 2 \tab 34 \tab 2 \tab 8 \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2192 \tab 75 \tab 25 \tab 1 \tab - \tab 3 \tab 1 \tab - \tab - \tab 6 \tab - \tab 1 \tab 1 \tab 27 \tab 3 \tab 5 \tab 2 \tab - \tab 1 \tab 3 \tab 1 \tab 1 \tab - \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   Total \tab 24379 \tab 818 \tab 234 \tab 15 \tab 6 \tab 40 \tab 4 \tab 20 \tab 11 \tab 41 \tab 12 \tab 17 \tab 13 \tab 930 \tab 21 \tab 74 \tab 22 \tab 3 \tab 11 \tab 11 \tab 5 \tab 8 \tab 3 \tab 4032 \tab 30730 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{relhh10}},
\code{\link{relhh11}},
\code{\link{relhh12}},
\code{\link{relhh13}},
\code{\link{relhh14}},
\code{\link{relhh6}},
\code{\link{relhh7}},
\code{\link{relhh8}},
\code{\link{relhh9}}
}
\concept{Household Members and Structure}
\keyword{variable}
