% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_590.R
\name{relhhd1}
\alias{relhhd1}
\title{Relation of 1st person to household head}
\source{
General Social Survey https://gss.norc.org
}
\description{
relhhd1
}
\details{
Question What is (PERSON)'s relationship to (HEAD OF HOUSEHOLD)?  SEE CODES BELOW ROSTER
}
\section{Values}{

\itemize{
\item \verb{[1]} head of household
\item \verb{[2]} spouse
\item \verb{[3]} partner, fiance(e), boyfriend, girlfriend, etc.
\item \verb{[4]} child, natural or adopted, stepchild
\item \verb{[5]} grandchild, great-grandchild
\item \verb{[6]} son-in-law, daughter-in-law
\item \verb{[7]} parent
\item \verb{[8]} parent-in-law
\item \verb{[9]} grandparent, great-grandparent
\item \verb{[10]} uncle or aunt
\item \verb{[11]} niece or nephew
\item \verb{[12]} cousin
\item \verb{[13]} brother, sister (full, half, or step sibling)
\item \verb{[14]} sibling-in-law
\item \verb{[15]} other relative (great-aunt, grandniece, etc.)
\item \verb{[16]} roommate, housemate
\item \verb{[17]} friend
\item \verb{[18]} child of non-relative
\item \verb{[19]} other non-relative (employee, boarder)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5416/vshow}{GSS Data Explorer page for \code{relhhd1}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab head of household \tab not available in this release \tab Total \cr
   1993 \tab - \tab 1606 \tab - \tab 1606 \cr
   1994 \tab - \tab 2992 \tab - \tab 2992 \cr
   1996 \tab - \tab 2904 \tab - \tab 2904 \cr
   1998 \tab - \tab 2832 \tab - \tab 2832 \cr
   2000 \tab - \tab 2817 \tab - \tab 2817 \cr
   2002 \tab - \tab 2765 \tab - \tab 2765 \cr
   2004 \tab - \tab 2812 \tab - \tab 2812 \cr
   2006 \tab - \tab 4510 \tab - \tab 4510 \cr
   2008 \tab - \tab 2023 \tab - \tab 2023 \cr
   2010 \tab - \tab 2044 \tab - \tab 2044 \cr
   2012 \tab - \tab 1974 \tab - \tab 1974 \cr
   2014 \tab - \tab 2538 \tab - \tab 2538 \cr
   2016 \tab - \tab 2867 \tab - \tab 2867 \cr
   2018 \tab - \tab 2348 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 3544 \tab - \tab - \tab 3544 \cr
   Total \tab 3544 \tab 37032 \tab 4032 \tab 44608 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away14}},
\code{\link{mar14}},
\code{\link{relhhd2}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd6}},
\code{\link{relhhd7}},
\code{\link{where14}}
}
\concept{Household Members and Structure}
\keyword{variable}
