% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_590.R
\name{relhhd2}
\alias{relhhd2}
\title{Relation of 2nd person to household head}
\source{
General Social Survey https://gss.norc.org
}
\description{
relhhd2
}
\details{
Question What is (PERSON)'s relationship to (HEAD OF HOUSEHOLD)?  SEE CODES BELOW ROSTER
}
\section{Values}{

\itemize{
\item \code{1} head of household
\item \code{2} spouse
\item \code{3} partner, fiance(e), boyfriend, girlfriend, etc.
\item \code{4} child, natural or adopted, stepchild
\item \code{5} grandchild, great-grandchild
\item \code{6} son-in-law, daughter-in-law
\item \code{7} parent
\item \code{8} parent-in-law
\item \code{9} grandparent, great-grandparent
\item \code{10} uncle or aunt
\item \code{11} niece or nephew
\item \code{12} cousin
\item \code{13} brother, sister (full, half, or step sibling)
\item \code{14} sibling-in-law
\item \code{15} other relative (great-aunt, grandniece, etc.)
\item \code{16} roommate, housemate
\item \code{17} friend
\item \code{18} child of non-relative
\item \code{19} other non-relative (employee, boarder)
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5417/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllllllllllllllllllll}{
   year \tab iap \tab brother, sister (full, half, or step sibling) \tab child, natural or adopted, stepchild \tab cousin \tab friend \tab grandchild, great-grandchild \tab niece or nephew \tab no answer \tab other non-relative (employee, boarder) \tab parent \tab partner, fiance(e), boyfriend, girlfriend, etc. \tab roommate, housemate \tab sibling-in-law \tab son-in-law, daughter-in-law \tab spouse \tab uncle or aunt \tab don't know \tab grandparent, great-grandparent \tab other relative (great-aunt, grandniece, etc.) \tab parent-in-law \tab child of non-relative \tab skipped on web \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 373 \tab 12 \tab 172 \tab 1 \tab 11 \tab 6 \tab 2 \tab 7 \tab 3 \tab 8 \tab 61 \tab 46 \tab 2 \tab 1 \tab 900 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 752 \tab 35 \tab 339 \tab 3 \tab 18 \tab 13 \tab 6 \tab 18 \tab 6 \tab 17 \tab 106 \tab 76 \tab 1 \tab 4 \tab 1596 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 743 \tab 21 \tab 353 \tab 4 \tab 18 \tab 16 \tab 5 \tab 89 \tab 4 \tab 20 \tab 144 \tab 82 \tab 3 \tab 3 \tab 1394 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 777 \tab 22 \tab 299 \tab 1 \tab 12 \tab 8 \tab 5 \tab 75 \tab 2 \tab 15 \tab 138 \tab 74 \tab 2 \tab 3 \tab 1389 \tab 3 \tab 1 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 741 \tab 29 \tab 353 \tab 3 \tab 13 \tab 13 \tab 9 \tab 10 \tab - \tab 26 \tab 192 \tab 92 \tab 1 \tab 4 \tab 1327 \tab 1 \tab - \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 883 \tab 36 \tab 328 \tab 4 \tab 24 \tab 7 \tab 1 \tab 5 \tab 19 \tab 11 \tab 163 \tab 51 \tab 2 \tab 13 \tab 1214 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2 \tab - \tab - \tab 2765 \cr
   2004 \tab 707 \tab 27 \tab 277 \tab 3 \tab 8 \tab 13 \tab 5 \tab 1 \tab 2 \tab 22 \tab 163 \tab 44 \tab 2 \tab 5 \tab 1527 \tab 1 \tab 3 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab 2812 \cr
   2006 \tab 1226 \tab 40 \tab 464 \tab 6 \tab 28 \tab 18 \tab 10 \tab 2 \tab 5 \tab 31 \tab 301 \tab 80 \tab 1 \tab 4 \tab 2288 \tab 1 \tab 2 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 527 \tab 25 \tab 228 \tab 4 \tab 10 \tab 15 \tab 4 \tab 1 \tab 3 \tab 11 \tab 151 \tab 31 \tab - \tab 5 \tab 1002 \tab 2 \tab 3 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 595 \tab 17 \tab 225 \tab - \tab 12 \tab 10 \tab 7 \tab 9 \tab 3 \tab 18 \tab 192 \tab 36 \tab - \tab 4 \tab 909 \tab 1 \tab 5 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 533 \tab 21 \tab 187 \tab 3 \tab 10 \tab 15 \tab 7 \tab 9 \tab 2 \tab 23 \tab 173 \tab 50 \tab 1 \tab 1 \tab 937 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 1974 \cr
   2014 \tab 678 \tab 26 \tab 290 \tab 3 \tab 7 \tab 13 \tab 7 \tab 6 \tab 3 \tab 17 \tab 218 \tab 53 \tab - \tab 8 \tab 1198 \tab 3 \tab 5 \tab 2 \tab - \tab - \tab 1 \tab - \tab - \tab 2538 \cr
   2016 \tab 849 \tab 39 \tab 315 \tab 3 \tab 14 \tab 16 \tab 5 \tab 3 \tab 1 \tab 22 \tab 269 \tab 67 \tab - \tab 3 \tab 1256 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 677 \tab 19 \tab 249 \tab 3 \tab 4 \tab 11 \tab 7 \tab 3 \tab 1 \tab 22 \tab 217 \tab 64 \tab 6 \tab 1 \tab 1057 \tab 1 \tab 3 \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 1845 \tab 2 \tab 1 \tab - \tab 7 \tab - \tab 1 \tab - \tab - \tab - \tab 314 \tab - \tab - \tab - \tab 1821 \tab - \tab - \tab - \tab - \tab - \tab - \tab 41 \tab - \tab 4032 \cr
   2022 \tab 2378 \tab 17 \tab 163 \tab 2 \tab 15 \tab 5 \tab 3 \tab 4 \tab 8 \tab 24 \tab 165 \tab 19 \tab 1 \tab 1 \tab 726 \tab 1 \tab - \tab 5 \tab - \tab 1 \tab 2 \tab 4 \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 42066 \tab 388 \tab 4243 \tab 43 \tab 211 \tab 179 \tab 84 \tab 242 \tab 62 \tab 287 \tab 2967 \tab 865 \tab 22 \tab 60 \tab 20541 \tab 23 \tab 27 \tab 17 \tab 4 \tab 7 \tab 7 \tab 45 \tab 3309 \tab 75699 \cr
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away14}},
\code{\link{mar14}},
\code{\link{relhhd1}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd6}},
\code{\link{relhhd7}},
\code{\link{where14}}

Other Cohabitation: 
\code{\link{relhhd1}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd6}},
\code{\link{relhhd7}}

Other Family: 
\code{\link{away14}},
\code{\link{mar14}},
\code{\link{relhhd1}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd6}},
\code{\link{relhhd7}},
\code{\link{where14}}

Other Household: 
\code{\link{away14}},
\code{\link{mar14}},
\code{\link{relhhd1}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd6}},
\code{\link{relhhd7}},
\code{\link{where14}}
}
\concept{Cohabitation}
\concept{Family}
\concept{Household}
\concept{Household Members and Structure}
\keyword{variable}
