% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_590.R
\name{relhhd6}
\alias{relhhd6}
\title{Relation of 6th person to household head}
\source{
General Social Survey https://gss.norc.org
}
\description{
relhhd6
}
\details{
Question What is (PERSON)'s relationship to (HEAD OF HOUSEHOLD)?  SEE CODES BELOW ROSTER
}
\section{Values}{

\itemize{
\item \code{1} head of household
\item \code{2} spouse
\item \code{3} partner, fiance(e), boyfriend, girlfriend, etc.
\item \code{4} child, natural or adopted, stepchild
\item \code{5} grandchild, great-grandchild
\item \code{6} son-in-law, daughter-in-law
\item \code{7} parent
\item \code{8} parent-in-law
\item \code{9} grandparent, great-grandparent
\item \code{10} uncle or aunt
\item \code{11} niece or nephew
\item \code{12} cousin
\item \code{13} brother, sister (full, half, or step sibling)
\item \code{14} sibling-in-law
\item \code{15} other relative (great-aunt, grandniece, etc.)
\item \code{16} roommate, housemate
\item \code{17} friend
\item \code{18} child of non-relative
\item \code{19} other non-relative (employee, boarder)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5421/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllllllllllllllllllll}{
   year \tab iap \tab child of non-relative \tab child, natural or adopted, stepchild \tab cousin \tab friend \tab grandchild, great-grandchild \tab niece or nephew \tab roommate, housemate \tab sibling-in-law \tab son-in-law, daughter-in-law \tab spouse \tab grandparent, great-grandparent \tab no answer \tab other non-relative (employee, boarder) \tab parent \tab parent-in-law \tab brother, sister (full, half, or step sibling) \tab other relative (great-aunt, grandniece, etc.) \tab don't know \tab uncle or aunt \tab partner, fiance(e), boyfriend, girlfriend, etc. \tab skipped on web \tab not available in this release \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1566 \tab 1 \tab 31 \tab 1 \tab 1 \tab 4 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2900 \tab 2 \tab 66 \tab - \tab - \tab 14 \tab 3 \tab 1 \tab 3 \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2791 \tab 3 \tab 72 \tab - \tab 1 \tab 17 \tab 4 \tab - \tab 2 \tab 4 \tab - \tab 1 \tab 5 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2741 \tab 4 \tab 63 \tab - \tab - \tab 7 \tab 5 \tab - \tab 1 \tab - \tab 1 \tab - \tab 5 \tab - \tab 2 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2721 \tab 1 \tab 68 \tab 1 \tab - \tab 8 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab 7 \tab - \tab 3 \tab - \tab - \tab - \tab 1 \tab 1 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2701 \tab 1 \tab 45 \tab - \tab 1 \tab 11 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2724 \tab 1 \tab 65 \tab 1 \tab - \tab 10 \tab 2 \tab 4 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4358 \tab 3 \tab 119 \tab 1 \tab 1 \tab 19 \tab 4 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 4510 \cr
   2008 \tab 1950 \tab 1 \tab 49 \tab - \tab 1 \tab 12 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1950 \tab 1 \tab 54 \tab - \tab 1 \tab 22 \tab 9 \tab 2 \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1898 \tab 1 \tab 56 \tab - \tab - \tab 8 \tab 4 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2463 \tab - \tab 58 \tab 1 \tab - \tab 7 \tab 3 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab 2538 \cr
   2016 \tab 2770 \tab 2 \tab 70 \tab - \tab - \tab 16 \tab 3 \tab 1 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2294 \tab - \tab 33 \tab - \tab - \tab 13 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab 1 \tab 1 \tab 1 \tab 2 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 3899 \tab 3 \tab 89 \tab 1 \tab 4 \tab 17 \tab 4 \tab 2 \tab 1 \tab 1 \tab - \tab - \tab - \tab 4 \tab 1 \tab 1 \tab 2 \tab - \tab - \tab 1 \tab - \tab 2 \tab - \tab 4032 \cr
   2022 \tab 3486 \tab - \tab 33 \tab - \tab 1 \tab 8 \tab 2 \tab - \tab - \tab 1 \tab - \tab 3 \tab - \tab 1 \tab 5 \tab 2 \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 70994 \tab 24 \tab 971 \tab 6 \tab 11 \tab 193 \tab 51 \tab 19 \tab 13 \tab 14 \tab 2 \tab 6 \tab 20 \tab 11 \tab 20 \tab 7 \tab 11 \tab 8 \tab 1 \tab 4 \tab 2 \tab 2 \tab 3309 \tab 75699 \cr
}
}

\seealso{
Other Household Members and Structure: 
\code{\link{away14}},
\code{\link{mar14}},
\code{\link{relhhd1}},
\code{\link{relhhd2}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd7}},
\code{\link{where14}}

Other Cohabitation: 
\code{\link{relhhd1}},
\code{\link{relhhd2}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd7}}

Other Family: 
\code{\link{away14}},
\code{\link{mar14}},
\code{\link{relhhd1}},
\code{\link{relhhd2}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd7}},
\code{\link{where14}}

Other Household: 
\code{\link{away14}},
\code{\link{mar14}},
\code{\link{relhhd1}},
\code{\link{relhhd2}},
\code{\link{relhhd3}},
\code{\link{relhhd4}},
\code{\link{relhhd5}},
\code{\link{relhhd7}},
\code{\link{where14}}
}
\concept{Cohabitation}
\concept{Family}
\concept{Household}
\concept{Household Members and Structure}
\keyword{variable}
