% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_193.R
\name{relhome}
\alias{relhome}
\title{Relationship to the previous owner of the home}
\source{
General Social Survey https://gss.norc.org
}
\description{
relhome
}
\details{
Question Which of the following best describes your relationship to the previous owner of the home you  purchased?
}
\section{Values}{

\itemize{
\item \code{1} a relative including in-laws
\item \code{2} a friend or acquaintance
\item \code{3} a friend of a friend-relative
\item \code{4} had previous business dealings
\item \code{5} no prior relationship
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1816/vshow}{GSS Data Explorer page for \code{relhome}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a relative including in-laws \tab a friend or acquaintance \tab a friend of a friend-relative \tab had previous business dealings \tab no prior relationship \tab no answer \tab Total \cr
   1996 \tab 2584 \tab 22 \tab 31 \tab 12 \tab 9 \tab 244 \tab 2 \tab 2904 \cr
   Total \tab 2584 \tab 22 \tab 31 \tab 12 \tab 9 \tab 244 \tab 2 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{evbuyhme}},
\code{\link{homeyear}},
\code{\link{learnhme}},
\code{\link{newowned}},
\code{\link{realtora}},
\code{\link{realtorb}},
\code{\link{realtorc}},
\code{\link{realtord}},
\code{\link{whosold}}
}
\concept{Markets}
\keyword{variable}
