% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_193.R
\name{relhome}
\alias{relhome}
\title{Relationship to the previous owner of the home}
\source{
General Social Survey https://gss.norc.org
}
\description{
relhome
}
\details{
Question Which of the following best describes your relationship to the previous owner of the home you  purchased?
}
\section{Values}{

\itemize{
\item \code{1} a relative including in-laws
\item \code{2} a friend or acquaintance
\item \code{3} a friend of a friend-relative
\item \code{4} had previous business dealings
\item \code{5} no prior relationship
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1816/vshow}{GSS Data Explorer page for \code{relhome}}.
}

\seealso{
Other Markets: 
\code{\link{evbuyhme}},
\code{\link{homeyear}},
\code{\link{learnhme}},
\code{\link{newowned}},
\code{\link{realtora}},
\code{\link{realtorb}},
\code{\link{realtorc}},
\code{\link{realtord}},
\code{\link{whosold}}

Other Purchases/Transactions: 
\code{\link{evbuyhme}},
\code{\link{homeyear}},
\code{\link{learnhme}},
\code{\link{newowned}},
\code{\link{realtora}},
\code{\link{realtorb}},
\code{\link{realtorc}},
\code{\link{realtord}},
\code{\link{whosold}}
}
\concept{Markets}
\concept{Purchases/Transactions}
\keyword{variable}
