% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_123.R
\name{relhrs2}
\alias{relhrs2}
\title{Hours religious act outside of home}
\source{
General Social Survey https://gss.norc.org
}
\description{
relhrs2
}
\details{
Question B. In the past month, about how many hours have you spent doing religious services activities outside your home (such as attending religious services, prayer groups, Bible studies, fellowship meetings, church leadership meetings, etc.)?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1144/vshow}{GSS Data Explorer page for \code{relhrs2}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 100 \tab 11 \tab 12 \tab 120 \tab 13 \tab 14 \tab 140 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 200 \tab 21 \tab 22 \tab 24 \tab 25 \tab 260 \tab 27 \tab 28 \tab 29 \tab 3 \tab 30 \tab 32 \tab 36 \tab 37 \tab 4 \tab 40 \tab 45 \tab 5 \tab 50 \tab 6 \tab 60 \tab 63 \tab 65 \tab 7 \tab 70 \tab 75 \tab 8 \tab 80 \tab 9 \tab 90 \tab 96 \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 749 \tab 53 \tab 64 \tab 1 \tab 1 \tab 26 \tab 1 \tab 1 \tab 6 \tab 1 \tab 30 \tab 13 \tab 1 \tab 6 \tab 1 \tab 61 \tab 38 \tab 3 \tab 1 \tab 1 \tab 7 \tab 6 \tab 1 \tab 1 \tab 2 \tab 1 \tab 42 \tab 19 \tab 4 \tab 1 \tab 1 \tab 98 \tab 9 \tab 1 \tab 40 \tab 7 \tab 38 \tab 7 \tab 1 \tab 2 \tab 9 \tab 2 \tab 2 \tab 46 \tab 1 \tab 5 \tab 2 \tab 1 \tab 5 \tab 26 \tab 2832 \cr
   Total \tab 1387 \tab 749 \tab 53 \tab 64 \tab 1 \tab 1 \tab 26 \tab 1 \tab 1 \tab 6 \tab 1 \tab 30 \tab 13 \tab 1 \tab 6 \tab 1 \tab 61 \tab 38 \tab 3 \tab 1 \tab 1 \tab 7 \tab 6 \tab 1 \tab 1 \tab 2 \tab 1 \tab 42 \tab 19 \tab 4 \tab 1 \tab 1 \tab 98 \tab 9 \tab 1 \tab 40 \tab 7 \tab 38 \tab 7 \tab 1 \tab 2 \tab 9 \tab 2 \tab 2 \tab 46 \tab 1 \tab 5 \tab 2 \tab 1 \tab 5 \tab 26 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{cathsch}},
\code{\link{givearts}},
\code{\link{giveoth}},
\code{\link{giverel}},
\code{\link{homesch}},
\code{\link{kid5up}},
\code{\link{numcong}},
\code{\link{pubsch}},
\code{\link{relhrs1}}
}
\concept{Religion}
\keyword{variable}
