% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_581.R
\name{relid2}
\alias{relid2}
\title{Protestant identification 2nd mention}
\source{
General Social Survey https://gss.norc.org
}
\description{
relid2
}
\details{
Question C. Do you consider yourself to be a member of (READ REMAINING GROUPS)?
}
\section{Values}{

\itemize{
\item \code{1} pentecostal
\item \code{2} fundamentalist
\item \code{3} evangelical
\item \code{4} mainline
\item \code{5} liberal
\item \code{6} none
\item \code{7} other (specify)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5327/vshow}{GSS Data Explorer page for \code{relid2}}.
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{attweek}},
\code{\link{cathid}},
\code{\link{charisma}},
\code{\link{othrel3}},
\code{\link{relid1}},
\code{\link{relid3}},
\code{\link{relidbst}},
\code{\link{religid}}
}
\concept{Religious Attendance and Identity}
\keyword{variable}
