% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_632.R
\name{relidesc}
\alias{relidesc}
\title{How well r's religion describes r}
\source{
General Social Survey https://gss.norc.org
}
\description{
relidesc
}
\details{
Question How well does being a member of your religion describe you?
}
\section{Values}{

\itemize{
\item \verb{[1]} extremely well
\item \verb{[2]} very well
\item \verb{[3]} somewhat well
\item \verb{[4]} not very well
\item \verb{[5]} not at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7527/vshow}{GSS Data Explorer page for \code{relidesc}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab extremely well \tab very well \tab somewhat well \tab not very well \tab not at all \tab don't know \tab no answer \tab skipped on web \tab see codebook \tab Total \cr
   2021 \tab 442 \tab 689 \tab 1085 \tab 1130 \tab 363 \tab 183 \tab 6 \tab 2 \tab 34 \tab 98 \tab 4032 \cr
   Total \tab 442 \tab 689 \tab 1085 \tab 1130 \tab 363 \tab 183 \tab 6 \tab 2 \tab 34 \tab 98 \tab 4032 \cr
}
}

\keyword{variable}
