% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_632.R
\name{relidimp}
\alias{relidimp}
\title{How important is r's religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
relidimp
}
\details{
Question How important is religion in your life – very important, somewhat important, not too important, or not at all important?
}
\section{Values}{

\itemize{
\item \verb{[1]} extremely important
\item \verb{[2]} very important
\item \verb{[3]} somewhat important
\item \verb{[4]} not very important
\item \verb{[5]} not important at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7526/vshow}{GSS Data Explorer page for \code{relidimp}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab extremely important \tab very important \tab somewhat important \tab not very important \tab not important at all \tab don't know \tab no answer \tab skipped on web \tab see codebook \tab Total \cr
   2021 \tab 390 \tab 855 \tab 776 \tab 711 \tab 563 \tab 606 \tab 3 \tab 2 \tab 28 \tab 98 \tab 4032 \cr
   Total \tab 390 \tab 855 \tab 776 \tab 711 \tab 563 \tab 606 \tab 3 \tab 2 \tab 28 \tab 98 \tab 4032 \cr
}
}

\keyword{variable}
