% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_632.R
\name{relidimp}
\alias{relidimp}
\title{How important is r's religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
relidimp
}
\details{
Question How important is religion in your life – very important, somewhat important, not too important, or not at all important?
}
\section{Values}{

\itemize{
\item \code{1} extremely important
\item \code{2} very important
\item \code{3} somewhat important
\item \code{4} not very important
\item \code{5} not important at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7526/vshow}{GSS Data Explorer page for \code{relidimp}}.
}

\keyword{variable}
