% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_632.R
\name{relidins}
\alias{relidins}
\title{To what extent r feels insulted if someone criticized r's religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
relidins
}
\details{
Question If someone criticized your religion, to what extent would it feel like a personal insult?
}
\section{Values}{

\itemize{
\item \verb{[1]} a great deal
\item \verb{[2]} somewhat
\item \verb{[3]} very little
\item \verb{[4]} not at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7529/vshow}{GSS Data Explorer page for \code{relidins}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab a great deal \tab somewhat \tab very little \tab not at all \tab don't know \tab no answer \tab skipped on web \tab see codebook \tab Total \cr
   2021 \tab 394 \tab 483 \tab 1115 \tab 927 \tab 971 \tab 3 \tab 4 \tab 37 \tab 98 \tab 4032 \cr
   Total \tab 394 \tab 483 \tab 1115 \tab 927 \tab 971 \tab 3 \tab 4 \tab 37 \tab 98 \tab 4032 \cr
}
}

\keyword{variable}
