% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_632.R
\name{relidins}
\alias{relidins}
\title{To what extent r feels insulted if someone criticized r's religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
relidins
}
\details{
Question If someone criticized your religion, to what extent would it feel like a personal insult?
}
\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} somewhat
\item \code{3} very little
\item \code{4} not at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7529/vshow}{GSS Data Explorer page for \code{relidins}}.
}

\keyword{variable}
