% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_632.R
\name{relidwe}
\alias{relidwe}
\title{How often r says 'we' when talking about r's religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
relidwe
}
\details{
Question When talking about your religion, how often do you say we instead of they?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} rarely
\item \code{3} some of the time
\item \code{4} most of the time
\item \code{5} all of the time
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7528/vshow}{GSS Data Explorer page for \code{relidwe}}.
}

\keyword{variable}
