% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_104.R
\name{relig2}
\alias{relig2}
\title{Religious preference of person #2}
\source{
General Social Survey https://gss.norc.org
}
\description{
relig2
}
\details{
Question What is (NAME 1)'s religious preference?  Is it Protestant, Catholic, Jewish, some other religion, or no religion?  PROBE:
What is your best guess?
ASK FOR EACH NAME REPEATING CATEGORIES AS NECESSARY.
2. Religion of Second Person
}
\section{Values}{

\itemize{
\item \code{1} protestant
\item \code{2} catholic
\item \code{3} jewish
\item \code{4} none
\item \code{5} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/951/vshow}{GSS Data Explorer page for \code{relig2}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab protestant \tab catholic \tab jewish \tab none \tab other \tab don't know \tab no answer \tab Total \cr
   1985 \tab 364 \tab 657 \tab 299 \tab 32 \tab 68 \tab 25 \tab 83 \tab 6 \tab 1534 \cr
   2004 \tab 1977 \tab 361 \tab 190 \tab 25 \tab 76 \tab 67 \tab 67 \tab 49 \tab 2812 \cr
   Total \tab 2341 \tab 1018 \tab 489 \tab 57 \tab 144 \tab 92 \tab 150 \tab 55 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{partyid1}},
\code{\link{partyid2}},
\code{\link{partyid3}},
\code{\link{relig1}},
\code{\link{relig3}},
\code{\link{relig4}},
\code{\link{relig5}},
\code{\link{talkpol1}},
\code{\link{talkpol2}}
}
\concept{Social Networks}
\keyword{variable}
