% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_104.R
\name{relig4}
\alias{relig4}
\title{Religious preference of person #4}
\source{
General Social Survey https://gss.norc.org
}
\description{
relig4
}
\details{
Question What is (NAME 1)'s religious preference?  Is it Protestant, Catholic, Jewish, some other religion, or no religion?  PROBE:
What is your best guess?
ASK FOR EACH NAME REPEATING CATEGORIES AS NECESSARY.
4. Religion of Fourth Person
}
\section{Values}{

\itemize{
\item \code{1} protestant
\item \code{2} catholic
\item \code{3} jewish
\item \code{4} none
\item \code{5} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/953/vshow}{GSS Data Explorer page for \code{relig4}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab protestant \tab catholic \tab jewish \tab none \tab other \tab don't know \tab no answer \tab Total \cr
   1985 \tab 920 \tab 343 \tab 159 \tab 18 \tab 37 \tab 12 \tab 41 \tab 4 \tab 1534 \cr
   2004 \tab 2472 \tab 134 \tab 68 \tab 8 \tab 29 \tab 20 \tab 33 \tab 48 \tab 2812 \cr
   Total \tab 3392 \tab 477 \tab 227 \tab 26 \tab 66 \tab 32 \tab 74 \tab 52 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{partyid1}},
\code{\link{partyid2}},
\code{\link{partyid3}},
\code{\link{relig1}},
\code{\link{relig2}},
\code{\link{relig3}},
\code{\link{relig5}},
\code{\link{talkpol1}},
\code{\link{talkpol2}}
}
\concept{Social Networks}
\keyword{variable}
