% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_663.R
\name{relig_next}
\alias{relig_next}
\title{Religion - follow-on version}
\source{
General Social Survey https://gss.norc.org
}
\description{
relig_next
}
\details{
Question What is your religious preference?
}
\section{Values}{

\itemize{
\item \code{1} protestant
\item \code{2} catholic
\item \code{3} jewish
\item \code{4} none
\item \code{5} other
\item \code{6} buddhism
\item \code{7} hinduism
\item \code{8} other eastern religions
\item \code{9} muslim/islam
\item \code{10} orthodox-christian
\item \code{11} christian
\item \code{12} native american
\item \code{13} inter-nondenominational
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8009/vshow}{GSS Data Explorer page for \code{relig_next}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab protestant \tab catholic \tab jewish \tab none \tab other \tab buddhism \tab hinduism \tab muslim/islam \tab orthodox-christian \tab christian \tab native american \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab - \tab 362 \tab 242 \tab 41 \tab 384 \tab 146 \tab 10 \tab 7 \tab 12 \tab 5 \tab 21 \tab 1 \tab 2308 \tab 5 \tab 3544 \cr
   Total \tab 0 \tab 362 \tab 242 \tab 41 \tab 384 \tab 146 \tab 10 \tab 7 \tab 12 \tab 5 \tab 21 \tab 1 \tab 2308 \tab 5 \tab 3544 \cr
}
}

\keyword{variable}
