% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_514.R
\name{religcon}
\alias{religcon}
\title{Religions bring more conflict}
\source{
General Social Survey https://gss.norc.org
}
\description{
religcon
}
\details{
Question Do you agree or disagree with the following statements?
A. Looking around the world, religions bring more conflict than peace.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4688/vshow}{GSS Data Explorer page for \code{religcon}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1548 \tab 101 \tab 324 \tab 257 \tab 407 \tab 114 \tab 39 \tab 42 \tab 2832 \cr
   2008 \tab 658 \tab 265 \tab 544 \tab 207 \tab 274 \tab 48 \tab 24 \tab 3 \tab 2023 \cr
   2018 \tab 1173 \tab 185 \tab 416 \tab 257 \tab 235 \tab 51 \tab 20 \tab 11 \tab 2348 \cr
   Total \tab 3379 \tab 551 \tab 1284 \tab 721 \tab 916 \tab 213 \tab 83 \tab 56 \tab 7203 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{antirel}},
\code{\link{befair}},
\code{\link{cantrust}},
\code{\link{godright}},
\code{\link{perright}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{socright}},
\code{\link{trustsci}}
}
\concept{ISSP Religion}
\keyword{variable}
