% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_581.R
\name{religid}
\alias{religid}
\title{R's religious identity}
\source{
General Social Survey https://gss.norc.org
}
\description{
religid
}
\details{
Question A. When it comes to your religious identity, would you say you are a fundamentalist, evangelical, mainline, or liberal Protestant or do none of these describe you?
}
\section{Values}{

\itemize{
\item \code{1} fundamentalist
\item \code{2} evangelical
\item \code{3} mainline
\item \code{4} liberal
\item \code{5} none
\item \code{6} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5325/vshow}{GSS Data Explorer page for \code{religid}}.
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{attweek}},
\code{\link{cathid}},
\code{\link{charisma}},
\code{\link{othrel3}},
\code{\link{relid1}},
\code{\link{relid2}},
\code{\link{relid3}},
\code{\link{relidbst}}
}
\concept{Religious Attendance and Identity}
\keyword{variable}
