% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_514.R
\name{religinf}
\alias{religinf}
\title{The u.s. would be a better country if religion had less influence}
\source{
General Social Survey https://gss.norc.org
}
\description{
religinf
}
\details{
Question Do you agree or disagree with the following statements?
C. The U.S. would be a better country if religion had less influence.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2021 \cr
   A/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4690/vshow}{GSS Data Explorer page for \code{religinf}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1998 \tab 1548 \tab 51 \tab 131 \tab 321 \tab 479 \tab 209 \tab 49 \tab 44 \tab - \tab 2832 \cr
   2021 \tab 427 \tab 514 \tab 565 \tab 1069 \tab 890 \tab 521 \tab 3 \tab 5 \tab 38 \tab 4032 \cr
   Total \tab 1975 \tab 565 \tab 696 \tab 1390 \tab 1369 \tab 730 \tab 52 \tab 49 \tab 38 \tab 6864 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{antirel}},
\code{\link{befair}},
\code{\link{cantrust}},
\code{\link{godright}},
\code{\link{perright}},
\code{\link{religcon}},
\code{\link{religint}},
\code{\link{socright}},
\code{\link{trustsci}}
}
\concept{ISSP Religion}
\keyword{variable}
