% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_514.R
\name{religint}
\alias{religint}
\title{Religious people are often too intolerant}
\source{
General Social Survey https://gss.norc.org
}
\description{
religint
}
\details{
Question Do you agree or disagree with the following statements?
B. People with very strong religious beliefs are often too intolerant of others.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
   2008 \tab B/C/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4689/vshow}{GSS Data Explorer page for \code{religint}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{antirel}},
\code{\link{befair}},
\code{\link{cantrust}},
\code{\link{godright}},
\code{\link{perright}},
\code{\link{religcon}},
\code{\link{religinf}},
\code{\link{socright}},
\code{\link{trustsci}}

Other Religion: 
\code{\link{godright}},
\code{\link{perright}},
\code{\link{religcon}},
\code{\link{religinf}},
\code{\link{socright}}
}
\concept{ISSP Religion}
\concept{Religion}
\keyword{variable}
