% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_512.R
\name{religkid}
\alias{religkid}
\title{Religion r was raised in}
\source{
General Social Survey https://gss.norc.org
}
\description{
religkid
}
\details{
Question What religion, if any, were you raised in? Was it Protestant, Catholic, Jewish, some other religion, or no religion?
}
\section{Values}{

\itemize{
\item \code{1} protestant
\item \code{2} catholic
\item \code{3} jewish
\item \code{4} none
\item \code{5} other
\item \code{6} buddhism
\item \code{7} hinduism
\item \code{8} other eastern religions
\item \code{9} muslim/islam
\item \code{10} orthodox-christian
\item \code{11} native american
\item \code{12} other christian
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2008 \tab B/C/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4669/vshow}{GSS Data Explorer page for \code{religkid}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{denkid}},
\code{\link{egomeans}},
\code{\link{madenkid}},
\code{\link{marelkid}},
\code{\link{ownfate}},
\code{\link{padenkid}},
\code{\link{parelkid}},
\code{\link{relexper}},
\code{\link{religsp}}

Other Religion: 
\code{\link{denkid}},
\code{\link{egomeans}},
\code{\link{madenkid}},
\code{\link{marelkid}},
\code{\link{ownfate}},
\code{\link{padenkid}},
\code{\link{parelkid}},
\code{\link{relexper}},
\code{\link{religsp}}
}
\concept{ISSP Religion}
\concept{Religion}
\keyword{variable}
