% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_35.R
\name{relitena}
\alias{relitena}
\title{Strength of affiliation}
\source{
General Social Survey https://gss.norc.org
}
\description{
relitena
}
\details{
Question C. Would you call yourself a strong Jew or not a very strong Jew?
}
\section{Values}{

\itemize{
\item \verb{[1]} strong
\item \verb{[2]} not very strong
\item \verb{[3]} somewhat strong
\item \verb{[4]} no religion
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/299/vshow}{GSS Data Explorer page for \code{relitena}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab not very strong \tab no answer \tab Total \cr
   2012 \tab 1954 \tab 12 \tab 8 \tab 1974 \cr
   Total \tab 1954 \tab 12 \tab 8 \tab 1974 \cr
}
}

\seealso{
Other Core: 
\code{\link{attend}},
\code{\link{fund}},
\code{\link{jew}},
\code{\link{jewaj}},
\code{\link{maattend}},
\code{\link{paattend}},
\code{\link{postlife}},
\code{\link{reliten}},
\code{\link{spattend}}
}
\concept{Core}
\keyword{variable}
