% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_640.R
\name{relitennv}
\alias{relitennv}
\title{Strength of affiliation (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
relitennv
}
\details{
Question Would you call yourself a strong (PREFERENCE NAMED IN RELIG) or a not very strong (PREFERENCE NAMED IN RELIG)?
}
\section{Values}{

\itemize{
\item \verb{[1]} strong
\item \verb{[2]} not very strong
\item \verb{[3]} somewhat strong (phone mode only: volunteered)
\item \verb{[4]} no religion
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7616/vshow}{GSS Data Explorer page for \code{relitennv}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strong \tab not very strong \tab somewhat strong (phone mode only: volunteered) \tab no religion \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 5 \tab 624 \tab 732 \tab 9 \tab 1121 \tab 5 \tab 1488 \tab 48 \tab 4032 \cr
   2022 \tab 234 \tab 102 \tab 134 \tab - \tab 439 \tab - \tab 2626 \tab 9 \tab 3544 \cr
   Total \tab 239 \tab 726 \tab 866 \tab 9 \tab 1560 \tab 5 \tab 4114 \tab 57 \tab 7576 \cr
}
}

\keyword{variable}
