% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_640.R
\name{relitennv}
\alias{relitennv}
\title{Strength of affiliation (no volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
relitennv
}
\details{
Question Would you call yourself a strong (PREFERENCE NAMED IN RELIG) or a not very strong (PREFERENCE NAMED IN RELIG)?
}
\section{Values}{

\itemize{
\item \code{1} strong
\item \code{2} not very strong
\item \code{3} somewhat strong (phone mode only: volunteered)
\item \code{4} no religion
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7616/vshow}{GSS Data Explorer page for \code{relitennv}}.
}

\keyword{variable}
