% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_195.R
\name{rellaw}
\alias{rellaw}
\title{Relationship to the attorney}
\source{
General Social Survey https://gss.norc.org
}
\description{
rellaw
}
\details{
Question What was your relationship to the attorney who provided the services, or to another attorney in the firm? Please indicate what your closest relationship was.
}
\section{Values}{

\itemize{
\item \code{1} a relative including in-laws
\item \code{2} a friend or acquaintance
\item \code{3} a friend of a friend or relative etc
\item \code{4} previous business dealings
\item \code{5} no prior relationship
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1837/vshow}{GSS Data Explorer page for \code{rellaw}}.
}

\seealso{
Other Markets: 
\code{\link{fixera}},
\code{\link{fixerb}},
\code{\link{fixhome}},
\code{\link{lawyer1}},
\code{\link{lawyerd}},
\code{\link{lawyere}},
\code{\link{lawyerf}},
\code{\link{lawyerg}},
\code{\link{satlawyr}}
}
\concept{Markets}
\keyword{variable}
