% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_228.R
\name{rellife}
\alias{rellife}
\title{Try to carry beliefs into other dealings}
\source{
General Social Survey https://gss.norc.org
}
\description{
rellife
}
\details{
Question I try hard to carry my religious beliefs over into all my other dealings in life.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998, 2006, 2008, 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2104/vshow}{GSS Data Explorer page for \code{rellife}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1387 \tab 391 \tab 632 \tab 288 \tab 103 \tab 8 \tab 23 \tab 2832 \cr
   2006 \tab 1518 \tab 862 \tab 1218 \tab 561 \tab 308 \tab 35 \tab 8 \tab 4510 \cr
   2008 \tab - \tab 575 \tab 823 \tab 403 \tab 188 \tab 22 \tab 12 \tab 2023 \cr
   2010 \tab - \tab 557 \tab 827 \tab 393 \tab 226 \tab 23 \tab 18 \tab 2044 \cr
   Total \tab 2905 \tab 2385 \tab 3500 \tab 1645 \tab 825 \tab 88 \tab 61 \tab 11409 \cr
}
}

\seealso{
Other Religion and Culture: 
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{feelgod}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{godwatch}},
\code{\link{lesspain}},
\code{\link{relcmfrt}}
}
\concept{Religion and Culture}
\keyword{variable}
