% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_538.R
\name{relmeet}
\alias{relmeet}
\title{Allowed to hold pub meeting for religious extremist}
\source{
General Social Survey https://gss.norc.org
}
\description{
relmeet
}
\details{
Question Should religious extremists be allowed to hold public meetings?
}
\section{Values}{

\itemize{
\item \code{1} should definitely be allowed
\item \code{2} should probably be allowed
\item \code{3} should probably not be allowed
\item \code{4} should definitely not be allowed
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4917/vshow}{GSS Data Explorer page for \code{relmeet}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab should definitely be allowed \tab should probably be allowed \tab should probably not be allowed \tab should definitely not be allowed \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 458 \tab 591 \tab 239 \tab 142 \tab 31 \tab 11 \tab 2812 \cr
   2014 \tab 1274 \tab 407 \tab 456 \tab 162 \tab 123 \tab 112 \tab 4 \tab 2538 \cr
   Total \tab 2614 \tab 865 \tab 1047 \tab 401 \tab 265 \tab 143 \tab 15 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{actassoc}},
\code{\link{buypol}},
\code{\link{helpusa}},
\code{\link{helpwrld}},
\code{\link{milserve}},
\code{\link{obeylaws}},
\code{\link{othreasn}},
\code{\link{paytaxes}},
\code{\link{watchgov}}
}
\concept{ISSP Citizenship}
\keyword{variable}
