% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_217.R
\name{relmhsp1}
\alias{relmhsp1}
\title{Patient was self}
\source{
General Social Survey https://gss.norc.org
}
\description{
relmhsp1
}
\details{
Question Was this a relative, a close friend, or just someone you didn't know very well? (CIRCLE ALL THAT APPLY) PROBE UNTIL "NO": ANYONE ELSE?
A. Respondent.
}
\section{Values}{

\itemize{
\item \code{1} self
\item \code{2} not self
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2059/vshow}{GSS Data Explorer page for \code{relmhsp1}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab self \tab not self \tab no answer \tab Total \cr
   1996 \tab 2531 \tab 9 \tab 355 \tab 9 \tab 2904 \cr
   Total \tab 2531 \tab 9 \tab 355 \tab 9 \tab 2904 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{brkdimp1}},
\code{\link{brkdimp2}},
\code{\link{brkdimp3}},
\code{\link{brkdoth}},
\code{\link{brkdsym1}},
\code{\link{brkdsym2}},
\code{\link{brkdsym3}},
\code{\link{knwmhosp}},
\code{\link{relmhsp2}}
}
\concept{Mental Health}
\keyword{variable}
