% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_217.R
\name{relmhsp2}
\alias{relmhsp2}
\title{Patient was immediate family}
\source{
General Social Survey https://gss.norc.org
}
\description{
relmhsp2
}
\details{
Question Was this a relative, a close friend, or just someone you didn't know very well? (CIRCLE ALL THAT APPLY) PROBE UNTIL "NO": ANYONE ELSE?
B. Immediate Family.
}
\section{Values}{

\itemize{
\item \code{1} imm. fam
\item \code{2} not imm. fam
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2060/vshow}{GSS Data Explorer page for \code{relmhsp2}}.
}

\seealso{
Other Mental Health: 
\code{\link{brkdimp1}},
\code{\link{brkdimp2}},
\code{\link{brkdimp3}},
\code{\link{brkdoth}},
\code{\link{brkdsym1}},
\code{\link{brkdsym2}},
\code{\link{brkdsym3}},
\code{\link{knwmhosp}},
\code{\link{relmhsp1}}

Other Mental Health: 
\code{\link{brkdimp1}},
\code{\link{brkdimp2}},
\code{\link{brkdimp3}},
\code{\link{brkdoth}},
\code{\link{brkdsym1}},
\code{\link{brkdsym2}},
\code{\link{brkdsym3}},
\code{\link{knwmhosp}},
\code{\link{relmhsp1}}
}
\concept{Mental Health}
\keyword{variable}
