% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_453.R
\name{relmost}
\alias{relmost}
\title{Adult relative r has most contact with}
\source{
General Social Survey https://gss.norc.org
}
\description{
relmost
}
\details{
Question Thinking of all the adult relatives, which one do you have the most contact with?
}
\section{Values}{

\itemize{
\item \code{1} grmother
\item \code{2} grfather
\item \code{3} grdaughter
\item \code{4} grandson
\item \code{5} aunt
\item \code{6} uncle
\item \code{7} mother-in-law
\item \code{8} father-in-law
\item \code{9} sister-in-law
\item \code{10} brother-in-law
\item \code{11} female relative
\item \code{12} male relative
\item \code{13} none
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1986 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4158/vshow}{GSS Data Explorer page for \code{relmost}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{inlawnum}},
\code{\link{marcohab}},
\code{\link{posslqy}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}},
\code{\link{unaunum}}

Other Family: 
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{inlawnum}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{relvisit}},
\code{\link{unaunum}}
}
\concept{Family}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
