% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_517.R
\name{relobjct}
\alias{relobjct}
\title{For religious reasons r has in home shrine, altar, or religious object on displa}
\source{
General Social Survey https://gss.norc.org
}
\description{
relobjct
}
\details{
Question For religious reasons do you have in your home a shrine, altar, or a religious object on display such as an icon, menorah, or crucifix?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2018 \cr
   B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4709/vshow}{GSS Data Explorer page for \code{relobjct}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2008 \tab 658 \tab 531 \tab 825 \tab 4 \tab 5 \tab 2023 \cr
   2018 \tab 1173 \tab 466 \tab 695 \tab 2 \tab 12 \tab 2348 \cr
   Total \tab 1831 \tab 997 \tab 1520 \tab 6 \tab 17 \tab 4371 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{comfort}},
\code{\link{makefrnd}},
\code{\link{paxhappy}},
\code{\link{relsprt}},
\code{\link{rightpeo}},
\code{\link{vistholy}}
}
\concept{ISSP Religion}
\keyword{variable}
