% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_191.R
\name{relprivt}
\alias{relprivt}
\title{Relationship to the seller of the vehicle}
\source{
General Social Survey https://gss.norc.org
}
\description{
relprivt
}
\details{
Question Which of the following best describes your relationship to the person who sold you the vehicle at the time of the purchase?
}
\section{Values}{

\itemize{
\item \verb{[1]} a relative,including in-laws
\item \verb{[2]} a friend or acquaintance
\item \verb{[3]} a friend of a friend-relative
\item \verb{[4]} previous business dealings
\item \verb{[5]} no prior relationship
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1797/vshow}{GSS Data Explorer page for \code{relprivt}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a relative,including in-laws \tab a friend or acquaintance \tab a friend of a friend-relative \tab previous business dealings \tab no prior relationship \tab no answer \tab Total \cr
   1996 \tab 2653 \tab 39 \tab 58 \tab 18 \tab 16 \tab 119 \tab 1 \tab 2904 \cr
   Total \tab 2653 \tab 39 \tab 58 \tab 18 \tab 16 \tab 119 \tab 1 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{cardealr}},
\code{\link{carprivt}},
\code{\link{newused}},
\code{\link{typdealr}},
\code{\link{warrntyc}}
}
\concept{Markets}
\keyword{variable}
