% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_513.R
\name{relscrpt}
\alias{relscrpt}
\title{R has read religious text outside of a service in the past year}
\source{
General Social Survey https://gss.norc.org
}
\description{
relscrpt
}
\details{
Question During the last 12 months, have you read or listened to the reading of any holy scripture such as the Bible, Buddhist sutra, Koran, Sruti, Torah, or other religious scripture, not counting any reading that happened during a worship service?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7278/vshow}{GSS Data Explorer page for \code{relscrpt}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab can't choose \tab no answer \tab Total \cr
   2018 \tab 1173 \tab 685 \tab 478 \tab 1 \tab 11 \tab 2348 \cr
   Total \tab 1173 \tab 685 \tab 478 \tab 1 \tab 11 \tab 2348 \cr
}
}

\keyword{variable}
