% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_517.R
\name{relsprt}
\alias{relsprt}
\title{Is r spiritual person}
\source{
General Social Survey https://gss.norc.org
}
\description{
relsprt
}
\details{
Question What best describes you:
}
\section{Values}{

\itemize{
\item \code{1} i follow a religion and consider myself to be a spiritual person interested in the sacred and the supernatural
\item \code{2} i follow a religion, but don't consider myself to be a spiritual person interested in the sacred and the supernat
\item \code{3} i don't follow a religion, but consider myself to be a spiritual person interested in the sacred and the supernat
\item \code{4} i don't follow a religion and don't consider myself to be a spiritual person interested in the sacred and the sup
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab B/C \tab partial \cr
   2018 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4711/vshow}{GSS Data Explorer page for \code{relsprt}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{comfort}},
\code{\link{makefrnd}},
\code{\link{paxhappy}},
\code{\link{relobjct}},
\code{\link{rightpeo}},
\code{\link{vistholy}}
}
\concept{ISSP Religion}
\keyword{variable}
