% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_344.R
\name{relsprt2}
\alias{relsprt2}
\title{Find comfort in religion or spirituality}
\source{
General Social Survey https://gss.norc.org
}
\description{
relsprt2
}
\details{
Question B. I find comfort in my religion or spirituality
}
\section{Values}{

\itemize{
\item \verb{[1]} many times a day
\item \verb{[2]} every day
\item \verb{[3]} most days
\item \verb{[4]} some days
\item \verb{[5]} once in a while
\item \verb{[6]} never or almost never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3086/vshow}{GSS Data Explorer page for \code{relsprt2}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab many times a day \tab every day \tab most days \tab some days \tab once in a while \tab never or almost never \tab don't know \tab no answer \tab Total \cr
   2004 \tab 1472 \tab 254 \tab 422 \tab 244 \tab 118 \tab 143 \tab 136 \tab 7 \tab 16 \tab 2812 \cr
   Total \tab 1472 \tab 254 \tab 422 \tab 244 \tab 118 \tab 143 \tab 136 \tab 7 \tab 16 \tab 2812 \cr
}
}

\seealso{
Other Spirituality and Religion: 
\code{\link{blessngs}},
\code{\link{closrgod}},
\code{\link{godclose}},
\code{\link{godguide}},
\code{\link{godhelp}},
\code{\link{godlvdir}},
\code{\link{godlvoth}},
\code{\link{joylifts}},
\code{\link{relsprt1}}
}
\concept{Spirituality and Religion}
\keyword{variable}
