% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_453.R
\name{reltime}
\alias{reltime}
\title{Travel time to most contacted rel}
\source{
General Social Survey https://gss.norc.org
}
\description{
reltime
}
\details{
Question Thinking of all the adult relatives, which one do you have the most contact with?
B. About how long would it take you to get to where your relative lives? Think of the time it usually takes door to door.
}
\section{Values}{

\itemize{
\item \code{1} lt 15 min
\item \code{2} 15-30 min
\item \code{3} 30 min-1 hr
\item \code{4} 1-2 hr
\item \code{5} 2-3 hr
\item \code{6} 3-5 hr
\item \code{7} 5-12 hr
\item \code{8} 12+ hr
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4160/vshow}{GSS Data Explorer page for \code{reltime}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab lt 15 min \tab 15-30 min \tab 30 min-1 hr \tab 1-2 hr \tab 2-3 hr \tab 3-5 hr \tab 5-12 hr \tab 12+ hr \tab no answer \tab Total \cr
   1986 \tab 172 \tab 416 \tab 269 \tab 157 \tab 108 \tab 57 \tab 68 \tab 105 \tab 109 \tab 9 \tab 1470 \cr
   Total \tab 172 \tab 416 \tab 269 \tab 157 \tab 108 \tab 57 \tab 68 \tab 105 \tab 109 \tab 9 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{inlawnum}},
\code{\link{marcohab}},
\code{\link{posslqy}},
\code{\link{relmost}},
\code{\link{relnum}},
\code{\link{relvisit}},
\code{\link{unaunum}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
