% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_684.R
\name{reltrad16}
\alias{reltrad16}
\title{Religious tradition age 16}
\source{
General Social Survey https://gss.norc.org
}
\description{
reltrad16
}
\details{
Question RELTRAD16 is constructed from variables RELIG16, DENOM16 and OTH16. The missing values coded as ".z: Variable-specific reserve code" represent OTH16 denominations that are not originally covered by the Steensland et al. (2000) article (i.e., OTHER values larger than 148). Please see Appendix F in the codebook for more information.
}
\section{Values}{

\itemize{
\item \verb{[1]} evangelical protestant
\item \verb{[2]} mainline protestant
\item \verb{[3]} black protestant
\item \verb{[4]} catholic
\item \verb{[5]} jewish
\item \verb{[6]} other faith
\item \verb{[7]} nonaffiliated
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   Full \tab 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8291/vshow}{GSS Data Explorer page for \code{reltrad16}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab evangelical protestant \tab mainline protestant \tab black protestant \tab catholic \tab jewish \tab other faith \tab nonaffiliated \tab don't know \tab no answer \tab see codebook \tab skipped on web \tab Total \cr
   1973 \tab 24 \tab 311 \tab 464 \tab 156 \tab 419 \tab 43 \tab 33 \tab 36 \tab 7 \tab 11 \tab - \tab - \tab 1504 \cr
   1974 \tab 22 \tab 323 \tab 464 \tab 154 \tab 402 \tab 49 \tab 21 \tab 39 \tab 6 \tab 4 \tab - \tab - \tab 1484 \cr
   1975 \tab 30 \tab 339 \tab 497 \tab 141 \tab 378 \tab 25 \tab 29 \tab 45 \tab 3 \tab 3 \tab - \tab - \tab 1490 \cr
   1976 \tab 27 \tab 350 \tab 467 \tab 110 \tab 428 \tab 33 \tab 34 \tab 47 \tab - \tab 3 \tab - \tab - \tab 1499 \cr
   1977 \tab 31 \tab 352 \tab 494 \tab 146 \tab 390 \tab 37 \tab 31 \tab 39 \tab 1 \tab 9 \tab - \tab - \tab 1530 \cr
   1978 \tab 33 \tab 362 \tab 453 \tab 145 \tab 418 \tab 31 \tab 33 \tab 48 \tab - \tab 9 \tab - \tab - \tab 1532 \cr
   1980 \tab 25 \tab 326 \tab 480 \tab 122 \tab 389 \tab 32 \tab 42 \tab 45 \tab - \tab 7 \tab - \tab - \tab 1468 \cr
   1982 \tab 33 \tab 316 \tab 511 \tab 414 \tab 441 \tab 44 \tab 42 \tab 48 \tab - \tab 11 \tab - \tab - \tab 1860 \cr
   1983 \tab 26 \tab 331 \tab 447 \tab 139 \tab 481 \tab 50 \tab 63 \tab 52 \tab - \tab 10 \tab - \tab - \tab 1599 \cr
   1984 \tab 36 \tab 386 \tab 346 \tab 138 \tab 399 \tab 30 \tab 63 \tab 43 \tab 1 \tab 31 \tab - \tab - \tab 1473 \cr
   1985 \tab 23 \tab 423 \tab 370 \tab 127 \tab 444 \tab 32 \tab 46 \tab 63 \tab - \tab 6 \tab - \tab - \tab 1534 \cr
   1986 \tab 26 \tab 386 \tab 338 \tab 149 \tab 412 \tab 33 \tab 71 \tab 45 \tab 3 \tab 6 \tab 1 \tab - \tab 1470 \cr
   1987 \tab 26 \tab 420 \tab 364 \tab 431 \tab 407 \tab 22 \tab 63 \tab 48 \tab 1 \tab 37 \tab - \tab - \tab 1819 \cr
   1988 \tab 30 \tab 368 \tab 364 \tab 146 \tab 415 \tab 31 \tab 68 \tab 53 \tab - \tab 6 \tab - \tab - \tab 1481 \cr
   1989 \tab 61 \tab 379 \tab 359 \tab 131 \tab 444 \tab 28 \tab 65 \tab 61 \tab - \tab 9 \tab - \tab - \tab 1537 \cr
   1990 \tab 40 \tab 328 \tab 352 \tab 120 \tab 361 \tab 27 \tab 59 \tab 72 \tab - \tab 13 \tab - \tab - \tab 1372 \cr
   1991 \tab 37 \tab 393 \tab 344 \tab 153 \tab 434 \tab 34 \tab 58 \tab 58 \tab - \tab 6 \tab - \tab - \tab 1517 \cr
   1993 \tab 37 \tab 465 \tab 358 \tab 142 \tab 422 \tab 35 \tab 60 \tab 71 \tab - \tab 16 \tab - \tab - \tab 1606 \cr
   1994 \tab 69 \tab 786 \tab 643 \tab 285 \tab 882 \tab 55 \tab 117 \tab 127 \tab 3 \tab 23 \tab 2 \tab - \tab 2992 \cr
   1996 \tab 60 \tab 694 \tab 633 \tab 281 \tab 837 \tab 68 \tab 141 \tab 169 \tab 2 \tab 18 \tab 1 \tab - \tab 2904 \cr
   1998 \tab 60 \tab 684 \tab 574 \tab 272 \tab 857 \tab 53 \tab 97 \tab 175 \tab 7 \tab 52 \tab 1 \tab - \tab 2832 \cr
   2000 \tab 91 \tab 655 \tab 515 \tab 304 \tab 844 \tab 64 \tab 119 \tab 192 \tab 5 \tab 28 \tab - \tab - \tab 2817 \cr
   2002 \tab 107 \tab 577 \tab 573 \tab 267 \tab 821 \tab 56 \tab 127 \tab 195 \tab 14 \tab 27 \tab 1 \tab - \tab 2765 \cr
   2004 \tab 150 \tab 622 \tab 502 \tab 254 \tab 829 \tab 60 \tab 137 \tab 226 \tab 13 \tab 14 \tab 5 \tab - \tab 2812 \cr
   2006 \tab 1618 \tab 639 \tab 510 \tab 273 \tab 960 \tab 56 \tab 168 \tab 249 \tab 27 \tab 7 \tab 3 \tab - \tab 4510 \cr
   2008 \tab 86 \tab 429 \tab 334 \tab 196 \tab 635 \tab 39 \tab 92 \tab 168 \tab 26 \tab 14 \tab 4 \tab - \tab 2023 \cr
   2010 \tab 143 \tab 415 \tab 352 \tab 183 \tab 647 \tab 33 \tab 89 \tab 150 \tab 18 \tab 11 \tab 3 \tab - \tab 2044 \cr
   2012 \tab 126 \tab 395 \tab 323 \tab 183 \tab 653 \tab 28 \tab 83 \tab 162 \tab 15 \tab 4 \tab 2 \tab - \tab 1974 \cr
   2014 \tab 156 \tab 494 \tab 375 \tab 229 \tab 872 \tab 41 \tab 135 \tab 206 \tab 15 \tab 13 \tab 2 \tab - \tab 2538 \cr
   2016 \tab 191 \tab 566 \tab 406 \tab 287 \tab 923 \tab 55 \tab 143 \tab 248 \tab 33 \tab 12 \tab 3 \tab - \tab 2867 \cr
   2018 \tab 151 \tab 457 \tab 339 \tab 218 \tab 732 \tab 42 \tab 124 \tab 241 \tab 29 \tab 12 \tab 3 \tab - \tab 2348 \cr
   2021 \tab 335 \tab 581 \tab 696 \tab 232 \tab 1290 \tab 87 \tab 212 \tab 499 \tab 58 \tab 4 \tab - \tab 38 \tab 4032 \cr
   2022 \tab 338 \tab 514 \tab 477 \tab 258 \tab 1159 \tab 55 \tab 217 \tab 404 \tab 78 \tab 18 \tab 5 \tab 21 \tab 3544 \cr
   2024 \tab 218 \tab 556 \tab 468 \tab 276 \tab 1003 \tab 47 \tab 218 \tab 434 \tab 51 \tab 12 \tab 4 \tab 22 \tab 3309 \cr
   Total \tab 4466 \tab 15622 \tab 15192 \tab 7062 \tab 21428 \tab 1455 \tab 3100 \tab 4758 \tab 416 \tab 466 \tab 40 \tab 81 \tab 74086 \cr
}
}

\seealso{
Other Religion: 
\code{\link{conorelgsp}},
\code{\link{reltrad}}

Other Religion: 
\code{\link{conorelgsp}},
\code{\link{reltrad}}
}
\concept{Religion}
\keyword{variable}
