% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_515.R
\name{reltruth}
\alias{reltruth}
\title{How much truth in religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
reltruth
}
\details{
Question Which of the following statements come closest to your own views:
}
\section{Values}{

\itemize{
\item \code{1} little truth any
\item \code{2} basic truth many
\item \code{3} truth in one
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
   2008 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4695/vshow}{GSS Data Explorer page for \code{reltruth}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{geomobil}},
\code{\link{relgrpeq}},
\code{\link{rspctrel}},
\code{\link{volwkchr}},
\code{\link{volwkoth}},
\code{\link{volwkpol}},
\code{\link{volwkrel}}

Other Religion: 
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{relgrpeq}},
\code{\link{rspctrel}}
}
\concept{ISSP Religion}
\concept{Religion}
\keyword{variable}
