% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_515.R
\name{reltruth}
\alias{reltruth}
\title{How much truth in religion}
\source{
General Social Survey https://gss.norc.org
}
\description{
reltruth
}
\details{
Question Which of the following statements come closest to your own views:
}
\section{Values}{

\itemize{
\item \verb{[1]} little truth any
\item \verb{[2]} basic truth many
\item \verb{[3]} truth in one
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4695/vshow}{GSS Data Explorer page for \code{reltruth}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab little truth any \tab basic truth many \tab truth in one \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1548 \tab 42 \tab 894 \tab 126 \tab 178 \tab 44 \tab 2832 \cr
   2008 \tab 658 \tab 60 \tab 1103 \tab 166 \tab 27 \tab 9 \tab 2023 \cr
   Total \tab 2206 \tab 102 \tab 1997 \tab 292 \tab 205 \tab 53 \tab 4855 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{geomobil}},
\code{\link{relgrpeq}},
\code{\link{rspctrel}},
\code{\link{volwkchr}},
\code{\link{volwkoth}},
\code{\link{volwkpol}},
\code{\link{volwkrel}}
}
\concept{ISSP Religion}
\keyword{variable}
