% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_453.R
\name{relvisit}
\alias{relvisit}
\title{How often visit most contacted rel}
\source{
General Social Survey https://gss.norc.org
}
\description{
relvisit
}
\details{
Question Thinking of all the adult relatives, which one do you have the most contact with?
A. How often do you see or visit with this relative?
}
\section{Values}{

\itemize{
\item \code{1} lives in hshld
\item \code{2} daily
\item \code{3} several x week
\item \code{4} once a week
\item \code{5} once a month
\item \code{6} several x year
\item \code{7} less often
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4159/vshow}{GSS Data Explorer page for \code{relvisit}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab lives in hshld \tab daily \tab several x week \tab once a week \tab once a month \tab several x year \tab less often \tab don't know \tab no answer \tab Total \cr
   1986 \tab 134 \tab 37 \tab 132 \tab 194 \tab 238 \tab 307 \tab 265 \tab 156 \tab 1 \tab 6 \tab 1470 \cr
   Total \tab 134 \tab 37 \tab 132 \tab 194 \tab 238 \tab 307 \tab 265 \tab 156 \tab 1 \tab 6 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{grkidnum}},
\code{\link{grparnum}},
\code{\link{inlawnum}},
\code{\link{marcohab}},
\code{\link{posslqy}},
\code{\link{relmost}},
\code{\link{relnum}},
\code{\link{reltime}},
\code{\link{unaunum}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
