% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_484.R
\name{repairs}
\alias{repairs}
\title{Who in household does small repair's}
\source{
General Social Survey https://gss.norc.org
}
\description{
repairs
}
\details{
Question In your household who does the following things . . .
B. Makes small repairs around the house.
}
\section{Values}{

\itemize{
\item \verb{[1]} always the woman
\item \verb{[2]} usually the woman
\item \verb{[3]} about equal or both together
\item \verb{[4]} usually the man
\item \verb{[5]} always the man
\item \verb{[6]} is done by a third person
\item \verb{[7]} not married
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4419/vshow}{GSS Data Explorer page for \code{repairs}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab always the woman \tab usually the woman \tab about equal or both together \tab usually the man \tab always the man \tab is done by a third person \tab not married \tab can't choose \tab no answer \tab Total \cr
   1994 \tab 1545 \tab 19 \tab 27 \tab 174 \tab 381 \tab 188 \tab 23 \tab 601 \tab 5 \tab 29 \tab 2992 \cr
   Total \tab 1545 \tab 19 \tab 27 \tab 174 \tab 381 \tab 188 \tab 23 \tab 601 \tab 5 \tab 29 \tab 2992 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{abchoose}},
\code{\link{caresick}},
\code{\link{chldcare}},
\code{\link{fambudgt}},
\code{\link{laundry}},
\code{\link{livnowed}},
\code{\link{mapaid}},
\code{\link{sexhar}},
\code{\link{teensex1}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
