% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_507.R
\name{replaceu}
\alias{replaceu}
\title{How easy r's firm replaces r if r left}
\source{
General Social Survey https://gss.norc.org
}
\description{
replaceu
}
\details{
Question B. How difficult or easy do you think it would be for your firm or organization to replace you if you left?
}
\section{Values}{

\itemize{
\item \code{1} very easy
\item \code{2} fairly easy
\item \code{3} neither easy nor difficult
\item \code{4} fairly difficult
\item \code{5} very difficult
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4619/vshow}{GSS Data Explorer page for \code{replaceu}}.
}

\seealso{
Other Work Orientation: 
\code{\link{dailywrk}},
\code{\link{famintjb}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobeasy}},
\code{\link{newjob}},
\code{\link{nounemp1}},
\code{\link{nounemp2}},
\code{\link{timeoff}}
}
\concept{Work Orientation}
\keyword{variable}
