% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_319.R
\name{reptileb4}
\alias{reptileb4}
\title{Reptile (turtle, snake, lizard)}
\source{
General Social Survey https://gss.norc.org
}
\description{
reptileb4
}
\details{
Question Past pet - Reptile
}
\section{Values}{

\itemize{
\item \code{0} did not have
\item \code{1} had
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7215/vshow}{GSS Data Explorer page for \code{reptileb4}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab did not have \tab had \tab no answer \tab Total \cr
   2018 \tab 2144 \tab 174 \tab 7 \tab 23 \tab 2348 \cr
   Total \tab 2144 \tab 174 \tab 7 \tab 23 \tab 2348 \cr
}
}

\keyword{variable}
