% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_199.R
\name{reqinfo}
\alias{reqinfo}
\title{Resp. of gov. to require biz. to provide info}
\source{
General Social Survey https://gss.norc.org
}
\description{
reqinfo
}
\details{
Question How much do you agree or disagree with the following statements:
B. It is the responsibility of government to require businesses to provide consumers with the information they need to make informed choices.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree somewhat
\item \verb{[3]} disagree somewhat
\item \verb{[4]} disagree strongly
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1872/vshow}{GSS Data Explorer page for \code{reqinfo}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab agree strongly \tab agree somewhat \tab disagree somewhat \tab disagree strongly \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 522 \tab 330 \tab 230 \tab 296 \tab 60 \tab 6 \tab 2904 \cr
   Total \tab 1460 \tab 522 \tab 330 \tab 230 \tab 296 \tab 60 \tab 6 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{decauto}},
\code{\link{econsys}},
\code{\link{fambiz}},
\code{\link{famfinan}},
\code{\link{haggle}},
\code{\link{hrfambiz}},
\code{\link{natrecon}},
\code{\link{tagsales}},
\code{\link{wkfambiz}}
}
\concept{Markets}
\keyword{variable}
