% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_533.R
\name{res161}
\alias{res161}
\title{Where spend childhood until age 16?}
\source{
General Social Survey https://gss.norc.org
}
\description{
res161
}
\details{
Question Where did you spend most of your childhood until age 16?
}
\section{Values}{

\itemize{
\item \verb{[1]} in this community
\item \verb{[2]} in a different community same state
\item \verb{[3]} in a different state in america
\item \verb{[4]} outside america
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4865/vshow}{GSS Data Explorer page for \code{res161}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab in this community \tab in a different community same state \tab in a different state in america \tab outside america \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 418 \tab 427 \tab 416 \tab 81 \tab 1 \tab 24 \tab 2904 \cr
   Total \tab 1537 \tab 418 \tab 427 \tab 416 \tab 81 \tab 1 \tab 24 \tab 2904 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{abroad}},
\code{\link{immameco}},
\code{\link{immideas}},
\code{\link{letin1}},
\code{\link{livecom1}},
\code{\link{meltpot1}},
\code{\link{refugees}}
}
\concept{ISSP National Identity}
\keyword{variable}
