% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_609.R
\name{res2006}
\alias{res2006}
\title{Was r living in usa during march-may 2006}
\source{
General Social Survey https://gss.norc.org
}
\description{
res2006
}
\details{
Question A. During March through May 2006, were you living in the United States or were you living outside of the country?
}
\section{Values}{

\itemize{
\item \verb{[1]} living in the u.s.
\item \verb{[2]} not living in u.s.
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5585/vshow}{GSS Data Explorer page for \code{res2006}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab living in the u.s. \tab not living in u.s. \tab no answer \tab don't know \tab Total \cr
   2008 \tab - \tab 1998 \tab 24 \tab 1 \tab - \tab 2023 \cr
   2010 \tab - \tab 2002 \tab 40 \tab - \tab 2 \tab 2044 \cr
   Total \tab 0 \tab 4000 \tab 64 \tab 1 \tab 2 \tab 4067 \cr
}
}

\seealso{
Other Sampling and Weights: 
\code{\link{cshutyp06}},
\code{\link{oversamp}},
\code{\link{phase}},
\code{\link{res2008}},
\code{\link{res2010}},
\code{\link{res2012}},
\code{\link{spaneng}},
\code{\link{spanint}},
\code{\link{spanself}}
}
\concept{Sampling and Weights}
\keyword{variable}
