% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_609.R
\name{res2008}
\alias{res2008}
\title{Was r living in usa during april-june 2008}
\source{
General Social Survey https://gss.norc.org
}
\description{
res2008
}
\details{
Question B. During April through June 2008, were you living in the United States or were you living outside of the country?
}
\section{Values}{

\itemize{
\item \code{1} living in the u.s.
\item \code{2} not living in u.s.
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5586/vshow}{GSS Data Explorer page for \code{res2008}}.
}

\seealso{
Other Sampling and Weights: 
\code{\link{ballot}},
\code{\link{batch}},
\code{\link{cshutyp06}},
\code{\link{cshutyp08}},
\code{\link{cshutyp10}},
\code{\link{cshutyp12}},
\code{\link{formwt}},
\code{\link{issp}},
\code{\link{oversamp}},
\code{\link{phase}},
\code{\link{res2006}},
\code{\link{res2010}},
\code{\link{res2012}},
\code{\link{sampcode}},
\code{\link{sample}()},
\code{\link{spaneng}},
\code{\link{spanint}},
\code{\link{spanself}},
\code{\link{subsamprate}},
\code{\link{version}},
\code{\link{vpsu}},
\code{\link{vstrat}},
\code{\link{wtss}},
\code{\link{wtssall}},
\code{\link{wtssnr}}
}
\concept{Sampling and Weights}
\keyword{variable}
