% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_609.R
\name{res2008}
\alias{res2008}
\title{Was r living in usa during april-june 2008}
\source{
General Social Survey https://gss.norc.org
}
\description{
res2008
}
\details{
Question B. During April through June 2008, were you living in the United States or were you living outside of the country?
}
\section{Values}{

\itemize{
\item \code{1} living in the u.s.
\item \code{2} not living in u.s.
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2010, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5586/vshow}{GSS Data Explorer page for \code{res2008}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab living in the u.s. \tab not living in u.s. \tab don't know \tab no answer \tab Total \cr
   2010 \tab - \tab 2023 \tab 20 \tab 1 \tab - \tab 2044 \cr
   2012 \tab 2 \tab 1931 \tab 40 \tab - \tab 1 \tab 1974 \cr
   Total \tab 2 \tab 3954 \tab 60 \tab 1 \tab 1 \tab 4018 \cr
}
}

\seealso{
Other Sampling and Weights: 
\code{\link{cshutyp06}},
\code{\link{oversamp}},
\code{\link{phase}},
\code{\link{res2006}},
\code{\link{res2010}},
\code{\link{res2012}},
\code{\link{spaneng}},
\code{\link{spanint}},
\code{\link{spanself}}
}
\concept{Sampling and Weights}
\keyword{variable}
