% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_609.R
\name{res2012}
\alias{res2012}
\title{Was r living in usa during april-june 2012}
\source{
General Social Survey https://gss.norc.org
}
\description{
res2012
}
\details{
Question D. During April through June 2012, were you living in the United States or were you living outside of the country?
}
\section{Values}{

\itemize{
\item \code{1} living in the u.s.
\item \code{2} not living in u.s.
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5588/vshow}{GSS Data Explorer page for \code{res2012}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab living in the u.s. \tab not living in u.s. \tab no answer \tab Total \cr
   2014 \tab - \tab 2498 \tab 39 \tab 1 \tab 2538 \cr
   Total \tab 0 \tab 2498 \tab 39 \tab 1 \tab 2538 \cr
}
}

\seealso{
Other Sampling and Weights: 
\code{\link{cshutyp06}},
\code{\link{oversamp}},
\code{\link{phase}},
\code{\link{res2006}},
\code{\link{res2008}},
\code{\link{res2010}},
\code{\link{spaneng}},
\code{\link{spanint}},
\code{\link{spanself}}
}
\concept{Sampling and Weights}
\keyword{variable}
