% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_306.R
\name{respect}
\alias{respect}
\title{R treated with respect at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
respect
}
\details{
Question Now I''m going to read you a list of statements that might or might not describe your main job. Please tell me whether you strongly agree, agree, disagree, or strongly disagree with each of these statements.
I. At the place where I work, I am treated with respect
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2782/vshow}{GSS Data Explorer page for \code{respect}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 715 \tab 924 \tab 105 \tab 31 \tab 6 \tab 15 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 696 \tab 880 \tab 97 \tab 36 \tab 10 \tab 15 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 438 \tab 628 \tab 72 \tab 23 \tab 2 \tab 24 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 442 \tab 704 \tab 77 \tab 21 \tab - \tab 6 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 478 \tab 815 \tab 101 \tab 18 \tab 3 \tab 33 \tab - \tab 2348 \cr
   2022 \tab 1546 \tab 766 \tab 1024 \tab 127 \tab 44 \tab 9 \tab 10 \tab 18 \tab 3544 \cr
   Total \tab 8336 \tab 3535 \tab 4975 \tab 579 \tab 173 \tab 30 \tab 103 \tab 18 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{knowwhat}},
\code{\link{lotofsay}},
\code{\link{myskills}},
\code{\link{overwork}},
\code{\link{safefrst}},
\code{\link{safetywk}},
\code{\link{teamsafe}},
\code{\link{trustman}},
\code{\link{wktopsat}}
}
\concept{Quality of Working Life}
\keyword{variable}
