% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_630.R
\name{respineq}
\alias{respineq}
\title{Who should have greatest responsibility for reducing inc differences}
\source{
General Social Survey https://gss.norc.org
}
\description{
respineq
}
\details{
Question From the following list, who do you think should have the greatest responsibility for reducing differences in income between people with high incomes and people with low incomes?
}
\section{Values}{

\itemize{
\item \code{1} private companies
\item \code{2} government
\item \code{3} trade unions
\item \code{4} high-income individuals themselves
\item \code{5} low-income individuals themselves
\item \code{6} income differences do not need to be reduced
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7506/vshow}{GSS Data Explorer page for \code{respineq}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab private companies \tab government \tab trade unions \tab high-income individuals themselves \tab low-income individuals themselves \tab income differences do not need to be reduced \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 2029 \tab 515 \tab 539 \tab 87 \tab 97 \tab 122 \tab 175 \tab 311 \tab 143 \tab 14 \tab 4032 \cr
   Total \tab 2029 \tab 515 \tab 539 \tab 87 \tab 97 \tab 122 \tab 175 \tab 311 \tab 143 \tab 14 \tab 4032 \cr
}
}

\keyword{variable}
