% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_630.R
\name{respineq}
\alias{respineq}
\title{Who should have greatest responsibility for reducing inc differences}
\source{
General Social Survey https://gss.norc.org
}
\description{
respineq
}
\details{
Question From the following list, who do you think should have the greatest responsibility for reducing differences in income between people with high incomes and people with low incomes?
}
\section{Values}{

\itemize{
\item \code{1} private companies
\item \code{2} government
\item \code{3} trade unions
\item \code{4} high-income individuals themselves
\item \code{5} low-income individuals themselves
\item \code{6} income differences do not need to be reduced
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7506/vshow}{GSS Data Explorer page for \code{respineq}}.
}

\keyword{variable}
