% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_591.R
\name{respnum}
\alias{respnum}
\title{Number in family of r}
\source{
General Social Survey https://gss.norc.org
}
\description{
respnum
}
\details{
Question Respondent's number in family.
}
\section{Values}{

\itemize{
\item \code{1} 1st person
\item \code{2} 2nd person
\item \code{3} 3rd person
\item \code{4} 4th person
\item \code{5} 5th person
\item \code{6} 6th person
\item \code{7} 7th person
\item \code{8} 8th person
\item \code{9} 9th person
\item \code{10} 10th person
\item \code{11} 11th person
\item \code{12} 12th person
\item \code{13} 13th person
\item \code{14} 14th person
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5432/vshow}{GSS Data Explorer page for \code{respnum}}.
}

\seealso{
Other Household Members and Structure: 
\code{\link{hefinfo}},
\code{\link{hhrace}},
\code{\link{relhhd10}},
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd8}},
\code{\link{relhhd9}}

Other Household: 
\code{\link{hefinfo}},
\code{\link{hhrace}},
\code{\link{relhhd10}},
\code{\link{relhhd11}},
\code{\link{relhhd12}},
\code{\link{relhhd13}},
\code{\link{relhhd14}},
\code{\link{relhhd8}},
\code{\link{relhhd9}}
}
\concept{Household}
\concept{Household Members and Structure}
\keyword{variable}
