% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_667.R
\name{respnumh}
\alias{respnumh}
\title{Number in family of r (hef)}
\source{
General Social Survey https://gss.norc.org
}
\description{
respnumh
}
\details{
Question Number in family of person who was respondent - selected from Household Enumeration Form
}
\section{Values}{

\itemize{
\item \code{1} 1st person
\item \code{2} 2nd person
\item \code{3} 3rd person
\item \code{4} 4th person
\item \code{5} 5th person
\item \code{6} 6th person
\item \code{7} 7th person
\item \code{8} 8th person
\item \code{9} 9th person
\item \code{10} 10th person
\item \code{11} 11th person
\item \code{12} 12th person
\item \code{13} 13th person
\item \code{14} 14th person
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8037/vshow}{GSS Data Explorer page for \code{respnumh}}.
}

\keyword{variable}
