% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_610.R
\name{respond}
\alias{respond}
\title{Estimate of the probability that this household will respond to the survey}
\source{
General Social Survey https://gss.norc.org
}
\description{
respond
}
\details{
Question Based on what you have learned about this housing unit, what is your estimate of the probability that this household will respond to the survey?
}
\section{Values}{

\itemize{
\item \code{1} high
\item \code{2} low
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6656/vshow}{GSS Data Explorer page for \code{respond}}.
}

\keyword{variable}
