% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_610.R
\name{respond}
\alias{respond}
\title{Estimate of the probability that this household will respond to the survey}
\source{
General Social Survey https://gss.norc.org
}
\description{
respond
}
\details{
Question Based on what you have learned about this housing unit, what is your estimate of the probability that this household will respond to the survey?
}
\section{Values}{

\itemize{
\item \code{1} high
\item \code{2} low
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2016, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6656/vshow}{GSS Data Explorer page for \code{respond}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab high \tab low \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2016 \tab - \tab 2174 \tab 661 \tab 5 \tab 27 \tab - \tab 2867 \cr
   2018 \tab - \tab 1752 \tab 576 \tab 14 \tab 6 \tab - \tab 2348 \cr
   2022 \tab 2077 \tab 952 \tab 515 \tab - \tab - \tab - \tab 3544 \cr
   Total \tab 2077 \tab 4878 \tab 1752 \tab 19 \tab 33 \tab 0 \tab 8759 \cr
}
}

\keyword{variable}
