% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_539.R
\name{revmeet}
\alias{revmeet}
\title{Allowed to hold pub meeting for ppl who want overthrow}
\source{
General Social Survey https://gss.norc.org
}
\description{
revmeet
}
\details{
Question Should people who want to overthrow the government by force be allowed to hold public meetings?
}
\section{Values}{

\itemize{
\item \code{1} should definitely be allowed
\item \code{2} should probably be allowed
\item \code{3} should probably not be allowed
\item \code{4} should definitely not be allowed
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4918/vshow}{GSS Data Explorer page for \code{revmeet}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab should definitely be allowed \tab should probably be allowed \tab should probably not be allowed \tab should definitely not be allowed \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 180 \tab 260 \tab 297 \tab 709 \tab 21 \tab 5 \tab 2812 \cr
   2014 \tab 1274 \tab 196 \tab 223 \tab 244 \tab 512 \tab 86 \tab 3 \tab 2538 \cr
   Total \tab 2614 \tab 376 \tab 483 \tab 541 \tab 1221 \tab 107 \tab 8 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{attrally}},
\code{\link{avoidbuy}},
\code{\link{cntctgov}},
\code{\link{interpol}},
\code{\link{joindem}},
\code{\link{polfunds}},
\code{\link{racmeet}},
\code{\link{signdpet}},
\code{\link{usemedia}}
}
\concept{ISSP Citizenship}
\keyword{variable}
