% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_444.R
\name{revoltrs}
\alias{revoltrs}
\title{Feel about people want to overthrow gov}
\source{
General Social Survey https://gss.norc.org
}
\description{
revoltrs
}
\details{
Question What are your personal feelings about...
B. People who want to overthrow the government by revolution.
}
\section{Values}{

\itemize{
\item \code{1} extremely favorable
\item \code{2} favorable
\item \code{3} neither favorable or unfavorable
\item \code{4} unfavorable
\item \code{5} extremely unfavorable
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4086/vshow}{GSS Data Explorer page for \code{revoltrs}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab extremely favorable \tab favorable \tab neither favorable or unfavorable \tab unfavorable \tab extremely unfavorable \tab can't choose \tab no answer \tab Total \cr
   1990 \tab 155 \tab 10 \tab 25 \tab 167 \tab 348 \tab 574 \tab 59 \tab 34 \tab 1372 \cr
   Total \tab 155 \tab 10 \tab 25 \tab 167 \tab 348 \tab 574 \tab 59 \tab 34 \tab 1372 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{aidcol}},
\code{\link{aidhouse}},
\code{\link{aidindus}},
\code{\link{aidold}},
\code{\link{aidunemp}},
\code{\link{equalize}},
\code{\link{hlthcare}},
\code{\link{protstrs}},
\code{\link{racists}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
