% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_435.R
\name{revtch15}
\alias{revtch15}
\title{Allow revolutionaries to teach children}
\source{
General Social Survey https://gss.norc.org
}
\description{
revtch15
}
\details{
Question There are some people whose views are considered extreme by the majority. First, consider people who want to overthrow the government by revolution. please circle a number after each question to show if you think such people should be allowed to . . .
B. Teach 15 year olds in schools.
}
\section{Values}{

\itemize{
\item \code{1} definitely allowed
\item \code{2} probably allowed
\item \code{3} probably not allowed
\item \code{4} definitely not allowed
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3989/vshow}{GSS Data Explorer page for \code{revtch15}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely allowed \tab probably allowed \tab probably not allowed \tab definitely not allowed \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 50 \tab 69 \tab 119 \tab 341 \tab 36 \tab 62 \tab 1534 \cr
   Total \tab 857 \tab 50 \tab 69 \tab 119 \tab 341 \tab 36 \tab 62 \tab 1534 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{crimhold}},
\code{\link{crimread}},
\code{\link{crimtail}},
\code{\link{crimtap}},
\code{\link{racpub}},
\code{\link{racspeak}},
\code{\link{ractch15}},
\code{\link{revpub}},
\code{\link{revspeak}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
