% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_474.R
\name{rewrdeff}
\alias{rewrdeff}
\title{Effort rewarded in america}
\source{
General Social Survey https://gss.norc.org
}
\description{
rewrdeff
}
\details{
Question Do you agree or disagree:
A. In America people get rewarded for their effort.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4320/vshow}{GSS Data Explorer page for \code{rewrdeff}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conage}},
\code{\link{firstjob}},
\code{\link{firstslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{occmobil}},
\code{\link{pajob}},
\code{\link{paslf}},
\code{\link{rewrdint}}

Other United States: 
\code{\link{rewrdint}}
}
\concept{ISSP Social Inequality}
\concept{United States}
\keyword{variable}
