% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_474.R
\name{rewrdint}
\alias{rewrdint}
\title{Intelligence and skill rewarded in america}
\source{
General Social Survey https://gss.norc.org
}
\description{
rewrdint
}
\details{
Question Do you agree or disagree:
B. In America people get rewarded for their intelligence and skills.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4321/vshow}{GSS Data Explorer page for \code{rewrdint}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2000 \tab 1545 \tab 187 \tab 696 \tab 208 \tab 90 \tab 17 \tab 21 \tab 53 \tab 2817 \cr
   Total \tab 1545 \tab 187 \tab 696 \tab 208 \tab 90 \tab 17 \tab 21 \tab 53 \tab 2817 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conage}},
\code{\link{firstjob}},
\code{\link{firstslf}},
\code{\link{lastjob}},
\code{\link{lastslf}},
\code{\link{occmobil}},
\code{\link{pajob}},
\code{\link{paslf}},
\code{\link{rewrdeff}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
