% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_294.R
\name{rfreenow}
\alias{rfreenow}
\title{R have more freedom than in past}
\source{
General Social Survey https://gss.norc.org
}
\description{
rfreenow
}
\details{
Question And how about you? Do you personally feel that you now have more freedom, less freedom, or about the same amount of freedom as you had in the past?
}
\section{Values}{

\itemize{
\item \code{1} more freedom
\item \code{2} about the same amount of freedom
\item \code{3} less freedom
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2666/vshow}{GSS Data Explorer page for \code{rfreenow}}.
}

\seealso{
Other Freedom: 
\code{\link{choice}},
\code{\link{freenow}},
\code{\link{howfree}},
\code{\link{inpeace}},
\code{\link{leftlone}},
\code{\link{nogovt}},
\code{\link{partpol}},
\code{\link{rhowfree}},
\code{\link{satdemoc}}

Other Freedom: 
\code{\link{choice}},
\code{\link{freenow}},
\code{\link{howfree}},
\code{\link{inpeace}},
\code{\link{leftlone}},
\code{\link{nogovt}},
\code{\link{partpol}},
\code{\link{rhowfree}},
\code{\link{satdemoc}}
}
\concept{Freedom}
\keyword{variable}
