% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_540.R
\name{rghtsmin}
\alias{rghtsmin}
\title{How important: govt protect right of minorities}
\source{
General Social Survey https://gss.norc.org
}
\description{
rghtsmin
}
\details{
Question There are different opinions about people's rights in a democracy. On a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it:
B. That government authorities respect and protect the rights of minorities
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all important
\item \verb{[7]} very important
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4934/vshow}{GSS Data Explorer page for \code{rghtsmin}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab not at all important \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 19 \tab 18 \tab 33 \tab 77 \tab 178 \tab 234 \tab 897 \tab 10 \tab 6 \tab 2812 \cr
   2014 \tab 1274 \tab 14 \tab 18 \tab 36 \tab 101 \tab 138 \tab 198 \tab 720 \tab 37 \tab 2 \tab 2538 \cr
   Total \tab 2614 \tab 33 \tab 36 \tab 69 \tab 178 \tab 316 \tab 432 \tab 1617 \tab 47 \tab 8 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{citviews}},
\code{\link{eqtreat}},
\code{\link{grpother}},
\code{\link{grpparty}},
\code{\link{grprelig}},
\code{\link{grpsprts}},
\code{\link{grpwork}},
\code{\link{polopts}},
\code{\link{solok}}
}
\concept{ISSP Citizenship}
\keyword{variable}
