% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_540.R
\name{rghtsmin}
\alias{rghtsmin}
\title{How important: govt protect right of minorities}
\source{
General Social Survey https://gss.norc.org
}
\description{
rghtsmin
}
\details{
Question There are different opinions about people's rights in a democracy. On a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it:
B. That government authorities respect and protect the rights of minorities
}
\section{Values}{

\itemize{
\item \code{1} not at all important
\item \code{7} very important
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4934/vshow}{GSS Data Explorer page for \code{rghtsmin}}.
}

\seealso{
Other ISSP Citizenship: 
\code{\link{citviews}},
\code{\link{eqtreat}},
\code{\link{grpother}},
\code{\link{grpparty}},
\code{\link{grprelig}},
\code{\link{grpsprts}},
\code{\link{grpwork}},
\code{\link{polopts}},
\code{\link{solok}}

Other Democracy: 
\code{\link{citviews}},
\code{\link{eqtreat}},
\code{\link{polopts}},
\code{\link{solok}}
}
\concept{Democracy}
\concept{ISSP Citizenship}
\keyword{variable}
