% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_475.R
\name{richeduc}
\alias{richeduc}
\title{Rich buy better education than poor}
\source{
General Social Survey https://gss.norc.org
}
\description{
richeduc
}
\details{
Question Is it just or unjust - right or wrong - that people with higher incomes can:
B. Buy better education for their children than people with lower incomes?
}
\section{Values}{

\itemize{
\item \code{1} very just, definitely right
\item \code{2} somewhat just, right
\item \code{3} neither just nor unjust, mixed feelings
\item \code{4} somewhat unjust, wrong
\item \code{5} very unjust, definitely wrong
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   B/C/- \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4327/vshow}{GSS Data Explorer page for \code{richeduc}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very just, definitely right \tab somewhat just, right \tab neither just nor unjust, mixed feelings \tab somewhat unjust, wrong \tab very unjust, definitely wrong \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2000 \tab 1545 \tab 178 \tab 200 \tab 242 \tab 237 \tab 339 \tab 26 \tab 50 \tab - \tab 2817 \cr
   2021 \tab 2029 \tab 119 \tab 241 \tab 422 \tab 404 \tab 583 \tab 70 \tab 146 \tab 18 \tab 4032 \cr
   Total \tab 3574 \tab 297 \tab 441 \tab 664 \tab 641 \tab 922 \tab 96 \tab 196 \tab 18 \tab 6849 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{corrupt}},
\code{\link{earndes}},
\code{\link{ldcgap}},
\code{\link{ldctax}},
\code{\link{payedtrn}},
\code{\link{payfam1}},
\code{\link{payresp}},
\code{\link{paysup}},
\code{\link{richhlth}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
