% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_475.R
\name{richhlth}
\alias{richhlth}
\title{Rich buy better health care than poor}
\source{
General Social Survey https://gss.norc.org
}
\description{
richhlth
}
\details{
Question Is it just or unjust - right or wrong - that people with higher incomes can:
A. Buy better health care than people with lower incomes
}
\section{Values}{

\itemize{
\item \code{1} very just, definitely right
\item \code{2} somewhat just, right
\item \code{3} neither just nor unjust, mixed feelings
\item \code{4} somewhat unjust, wrong
\item \code{5} very unjust, definitely wrong
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2021 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4326/vshow}{GSS Data Explorer page for \code{richhlth}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{corrupt}},
\code{\link{earndes}},
\code{\link{ldcgap}},
\code{\link{ldctax}},
\code{\link{payedtrn}},
\code{\link{payfam1}},
\code{\link{payresp}},
\code{\link{paysup}},
\code{\link{richeduc}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
