% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_475.R
\name{richhlth}
\alias{richhlth}
\title{Rich buy better health care than poor}
\source{
General Social Survey https://gss.norc.org
}
\description{
richhlth
}
\details{
Question Is it just or unjust - right or wrong - that people with higher incomes can:
A. Buy better health care than people with lower incomes
}
\section{Values}{

\itemize{
\item \verb{[1]} very just, definitely right
\item \verb{[2]} somewhat just, right
\item \verb{[3]} neither just nor unjust, mixed feelings
\item \verb{[4]} somewhat unjust, wrong
\item \verb{[5]} very unjust, definitely wrong
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   B/C/- \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4326/vshow}{GSS Data Explorer page for \code{richhlth}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab very just, definitely right \tab somewhat just, right \tab neither just nor unjust, mixed feelings \tab somewhat unjust, wrong \tab very unjust, definitely wrong \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2000 \tab 1545 \tab 165 \tab 170 \tab 264 \tab 249 \tab 340 \tab 35 \tab 49 \tab - \tab 2817 \cr
   2021 \tab 2029 \tab 121 \tab 191 \tab 452 \tab 421 \tab 586 \tab 73 \tab 144 \tab 15 \tab 4032 \cr
   Total \tab 3574 \tab 286 \tab 361 \tab 716 \tab 670 \tab 926 \tab 108 \tab 193 \tab 15 \tab 6849 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{corrupt}},
\code{\link{earndes}},
\code{\link{ldcgap}},
\code{\link{ldctax}},
\code{\link{payedtrn}},
\code{\link{payfam1}},
\code{\link{payresp}},
\code{\link{paysup}},
\code{\link{richeduc}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
