% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_406.R
\name{rifles50}
\alias{rifles50}
\title{Should high power rifle sale be limited to military}
\source{
General Social Survey https://gss.norc.org
}
\description{
rifles50
}
\details{
Question Currently under federal law, very high power, 50-caliber rifles that can penetrate armor from a mile away are available to people on the same basis as standard hunting rifles. Should such very high power rifles be 1) restricted Only to the police and military, or 2) available to civilians like standard hunting rifles as they are at present?
}
\section{Values}{

\itemize{
\item \code{1} restricted to the police and military
\item \code{2} available to civilians like other hunting rifles
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3704/vshow}{GSS Data Explorer page for \code{rifles50}}.
}

\seealso{
Other Gun Control: 
\code{\link{guns911}},
\code{\link{gunsales}},
\code{\link{gunsdrnk}},
\code{\link{gunsdrug}},
\code{\link{othguns}},
\code{\link{semiguns}}
}
\concept{Gun Control}
\keyword{variable}
