% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_517.R
\name{rightpeo}
\alias{rightpeo}
\title{Religion helps people to meet the right kind of people}
\source{
General Social Survey https://gss.norc.org
}
\description{
rightpeo
}
\details{
Question Do you agree or disagree that…
D. ...that practicing a religion helps people to meet the right kind of people?
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4715/vshow}{GSS Data Explorer page for \code{rightpeo}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{comfort}},
\code{\link{makefrnd}},
\code{\link{paxhappy}},
\code{\link{relobjct}},
\code{\link{relsprt}},
\code{\link{vistholy}}
}
\concept{ISSP Religion}
\keyword{variable}
