% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_250.R
\name{rightref}
\alias{rightref}
\title{Have a right to refuse further treatment}
\source{
General Social Survey https://gss.norc.org
}
\description{
rightref
}
\details{
Question How much do you agree of disagree with the following options for this person?  This person should...
A. Have a right to refuse further treatment except for pain relief.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2255/vshow}{GSS Data Explorer page for \code{rightref}}.
}

\seealso{
Other Mental Health: 
\code{\link{docsui}},
\code{\link{hospice}},
\code{\link{instrctn}},
\code{\link{notreat}},
\code{\link{painmed}},
\code{\link{termvig}},
\code{\link{trustcrt}},
\code{\link{trustdoc}},
\code{\link{trustfam}}

Other Terminal Illness: 
\code{\link{docsui}},
\code{\link{hospice}},
\code{\link{instrctn}},
\code{\link{notreat}},
\code{\link{painmed}},
\code{\link{termvig}},
\code{\link{trustcrt}},
\code{\link{trustdoc}},
\code{\link{trustfam}}
}
\concept{Mental Health}
\concept{Terminal Illness}
\keyword{variable}
