% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_176.R
\name{rimmdisc}
\alias{rimmdisc}
\title{Chances illegal immigrants favored over family}
\source{
General Social Survey https://gss.norc.org
}
\description{
rimmdisc
}
\details{
Question B. What about immigrants?  Is it very likely, somewhat likely, somewhat unlikely, or very unlikely that you or anyone in your family won't get a job or promotion while an equally or less qualified immigrant employee receives one instead?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} somewhat unlikely
\item \code{4} very unlikely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1648/vshow}{GSS Data Explorer page for \code{rimmdisc}}.
}

\seealso{
Other Multiculturalism: 
\code{\link{econfutr}},
\code{\link{futrdown}},
\code{\link{futrup}},
\code{\link{pastdown}},
\code{\link{pastup}},
\code{\link{rdiscaff}}
}
\concept{Multiculturalism}
\keyword{variable}
