% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_506.R
\name{rimpskls}
\alias{rimpskls}
\title{Job gives r chance to improve skills}
\source{
General Social Survey https://gss.norc.org
}
\description{
rimpskls
}
\details{
Question For each of these statements about your (main) job, please tell me how much you agree or disagree with it as it applies to your job.
A. My job gives me a chance to improve my skills
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4611/vshow}{GSS Data Explorer page for \code{rimpskls}}.
}

\seealso{
Other Work Orientation: 
\code{\link{adforjob}},
\code{\link{appemps}},
\code{\link{askhelp}},
\code{\link{econsup}},
\code{\link{selfemp1}},
\code{\link{selfemp2}},
\code{\link{unbetter}},
\code{\link{unjobsec}},
\code{\link{wantads}}

Other Work: 
\code{\link{selfemp1}},
\code{\link{selfemp2}},
\code{\link{unbetter}},
\code{\link{unjobsec}}
}
\concept{Work}
\concept{Work Orientation}
\keyword{variable}
