% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_16.R
\name{rincom77}
\alias{rincom77}
\title{Respondents income}
\source{
General Social Survey https://gss.norc.org
}
\description{
rincom77
}
\details{
Question Respondent's income on 1977/78/80 surveys. For question wording, see RINCOME.
}
\section{Values}{

\itemize{
\item \code{1} lt $1000
\item \code{2} $1000-2999
\item \code{3} $3000-3999
\item \code{4} $4000-4999
\item \code{5} $5000-5999
\item \code{6} $6000-6999
\item \code{7} $7000-7999
\item \code{8} $8000-9999
\item \code{9} $10000-12499
\item \code{10} $12500-14999
\item \code{11} $15000-17499
\item \code{12} $17500-19999
\item \code{13} $20000-22499
\item \code{14} $22500-24999
\item \code{15} $25000-49999
\item \code{16} $50000 +
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977, 1978, 1980 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/108/vshow}{GSS Data Explorer page for \code{rincom77}}.

Counts by year:\tabular{llllllllllllllllllllll}{
   year \tab iap \tab lt $1000 \tab $1000-2999 \tab $3000-3999 \tab $4000-4999 \tab $5000-5999 \tab $6000-6999 \tab $7000-7999 \tab $8000-9999 \tab $10000-12499 \tab $12500-14999 \tab $15000-17499 \tab $17500-19999 \tab $20000-22499 \tab $22500-24999 \tab $25000-49999 \tab $50000 + \tab don't know \tab no answer \tab refused \tab Total \cr
   1977 \tab 493 \tab 60 \tab 86 \tab 67 \tab 45 \tab 64 \tab 59 \tab 57 \tab 103 \tab 119 \tab 96 \tab 69 \tab 41 \tab 23 \tab 17 \tab 36 \tab 13 \tab 8 \tab 54 \tab 20 \tab 1530 \cr
   1978 \tab 551 \tab 70 \tab 113 \tab 48 \tab 53 \tab 52 \tab 42 \tab 44 \tab 80 \tab 122 \tab 73 \tab 71 \tab 52 \tab 34 \tab 26 \tab 50 \tab 7 \tab 12 \tab 8 \tab 24 \tab 1532 \cr
   1980 \tab 496 \tab 45 \tab 76 \tab 37 \tab 29 \tab 45 \tab 31 \tab 36 \tab 88 \tab 107 \tab 72 \tab 69 \tab 67 \tab 48 \tab 32 \tab 78 \tab 22 \tab 8 \tab 59 \tab 23 \tab 1468 \cr
   Total \tab 1540 \tab 175 \tab 275 \tab 152 \tab 127 \tab 161 \tab 132 \tab 137 \tab 271 \tab 348 \tab 241 \tab 209 \tab 160 \tab 105 \tab 75 \tab 164 \tab 42 \tab 28 \tab 121 \tab 67 \tab 4530 \cr
}
}

\seealso{
Other Core: 
\code{\link{income72}},
\code{\link{income77}},
\code{\link{income82}},
\code{\link{income86}},
\code{\link{income91}},
\code{\link{income98}},
\code{\link{rincom82}},
\code{\link{rincom86}},
\code{\link{rincom91}}
}
\concept{Core}
\keyword{variable}
