% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_16.R
\name{rincom77}
\alias{rincom77}
\title{Respondents income}
\source{
General Social Survey https://gss.norc.org
}
\description{
rincom77
}
\details{
Question Respondent's income on 1977/78/80 surveys. For question wording, see RINCOME.
}
\section{Values}{

\itemize{
\item \code{1} lt $1000
\item \code{2} $1000-2999
\item \code{3} $3000-3999
\item \code{4} $4000-4999
\item \code{5} $5000-5999
\item \code{6} $6000-6999
\item \code{7} $7000-7999
\item \code{8} $8000-9999
\item \code{9} $10000-12499
\item \code{10} $12500-14999
\item \code{11} $15000-17499
\item \code{12} $17500-19999
\item \code{13} $20000-22499
\item \code{14} $22500-24999
\item \code{15} $25000-49999
\item \code{16} $50000 +
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1977 \tab - \tab full \cr
   1978 \tab - \tab full \cr
   1980 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/108/vshow}{GSS Data Explorer page for \code{rincom77}}.
}

\seealso{
Other Core: 
\code{\link{income72}},
\code{\link{income77}},
\code{\link{income82}},
\code{\link{income86}},
\code{\link{income91}},
\code{\link{income98}},
\code{\link{rincom82}},
\code{\link{rincom86}},
\code{\link{rincom91}}
}
\concept{Core}
\keyword{variable}
