% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_500.R
\name{rleisure}
\alias{rleisure}
\title{R's job leaves much leisure time}
\source{
General Social Survey https://gss.norc.org
}
\description{
rleisure
}
\details{
Question For each statement about your main job below, please circle one code to show how much you agree or disagree that it applies to your job.
D. My job leaves a lot of leisure time
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4544/vshow}{GSS Data Explorer page for \code{rleisure}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 639 \tab 32 \tab 206 \tab 251 \tab 298 \tab 79 \tab 7 \tab 25 \tab 1537 \cr
   Total \tab 639 \tab 32 \tab 206 \tab 251 \tab 298 \tab 79 \tab 7 \tab 25 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{physwrk}},
\code{\link{rflexhrs}},
\code{\link{rhiinc}},
\code{\link{rhlpoths}},
\code{\link{rhlpsoc}},
\code{\link{rintjob}},
\code{\link{rpromotn}},
\code{\link{rwrkindp}},
\code{\link{xhaustn}}
}
\concept{Work Orientation}
\keyword{variable}
