% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_645.R
\name{rlyrltv}
\alias{rlyrltv}
\title{Rely on friends or close relatives}
\source{
General Social Survey https://gss.norc.org
}
\description{
rlyrltv
}
\details{
Question People have different opinions about who you can rely on and who should be important to you. While some people think that you can rely more on close relatives, others think that you can rely more on close friends. Which view is closer to your opinion?
}
\section{Values}{

\itemize{
\item \code{1} you can definitely rely more on close relatives
\item \code{4} you can rely equally on both
\item \code{7} you can definitely rely more on close friends
\item \code{8} you cannot rely on either of them
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7781/vshow}{GSS Data Explorer page for \code{rlyrltv}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab you can definitely rely more on close relatives \tab 2 \tab 3 \tab you can rely equally on both \tab 5 \tab 6 \tab you can definitely rely more on close friends \tab 9 \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2386 \tab 215 \tab 111 \tab 101 \tab 408 \tab 43 \tab 30 \tab 87 \tab 91 \tab 56 \tab 11 \tab 5 \tab 3544 \cr
   Total \tab 2386 \tab 215 \tab 111 \tab 101 \tab 408 \tab 43 \tab 30 \tab 87 \tab 91 \tab 56 \tab 11 \tab 5 \tab 3544 \cr
}
}

\keyword{variable}
