% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_338.R
\name{robject}
\alias{robject}
\title{Did r object to or delay answering card c4}
\source{
General Social Survey https://gss.norc.org
}
\description{
robject
}
\details{
Question INTERVIEWER: DID THE RESPONDENT OBJECT TO OR DELAY ANSWERING IMMSTATS?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, objected verbally
\item \verb{[2]} yes, delayed, kept silent for a while
\item \verb{[3]} no
\item \verb{[4]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3029/vshow}{GSS Data Explorer page for \code{robject}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes, objected verbally \tab yes, delayed, kept silent for a while \tab no \tab other \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2540 \tab 5 \tab 4 \tab 248 \tab 3 \tab 7 \tab 5 \tab 2812 \cr
   Total \tab 2540 \tab 5 \tab 4 \tab 248 \tab 3 \tab 7 \tab 5 \tab 2812 \cr
}
}

\seealso{
Other Immigration: 
\code{\link{bornhome}},
\code{\link{immstats}},
\code{\link{natborn}},
\code{\link{rundrstd}},
\code{\link{travelus}}
}
\concept{Immigration}
\keyword{variable}
