% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_558.R
\name{roleclrg}
\alias{roleclrg}
\title{Role of clergy with kids}
\source{
General Social Survey https://gss.norc.org
}
\description{
roleclrg
}
\details{
Question Many adults play roles in the lives of children. Please rate the following as very good, good, fairly good, fairly bad, or very bad. F. Priests, ministers and rabbis
}
\section{Values}{

\itemize{
\item \code{1} very good
\item \code{2} good
\item \code{3} fairly good
\item \code{4} fairly bad
\item \code{5} very bad
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5107/vshow}{GSS Data Explorer page for \code{roleclrg}}.
}

\seealso{
Other Children: 
\code{\link{chldeduc}},
\code{\link{chldhlth}},
\code{\link{chldhome}},
\code{\link{chldlove}},
\code{\link{chldmorl}},
\code{\link{chldsafe}},
\code{\link{chldskls}},
\code{\link{chldsup}},
\code{\link{chldtime}},
\code{\link{infadstv}},
\code{\link{inffilms}},
\code{\link{infmusic}},
\code{\link{infnettv}},
\code{\link{infpubtv}},
\code{\link{parfin}},
\code{\link{pargovt}},
\code{\link{parhardr}},
\code{\link{parrght}},
\code{\link{partaxes}},
\code{\link{partime}},
\code{\link{parwhere}},
\code{\link{parwork}},
\code{\link{parworse}},
\code{\link{roleccp}},
\code{\link{rolegp}},
\code{\link{rolema}},
\code{\link{rolepa}},
\code{\link{roletchr}},
\code{\link{spdrugs}},
\code{\link{spdsabkd}},
\code{\link{spfoodkd}},
\code{\link{spheadst}},
\code{\link{sphlthkd}},
\code{\link{sphomekd}},
\code{\link{sppill}},
\code{\link{sppoorkd}},
\code{\link{sppregnt}},
\code{\link{spwrkpar}}
}
\concept{Children}
\keyword{variable}
