% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_558.R
\name{roletchr}
\alias{roletchr}
\title{Role of teachers with kids}
\source{
General Social Survey https://gss.norc.org
}
\description{
roletchr
}
\details{
Question Many adults play roles in the lives of children. Please rate the following as very good, good, fairly good, fairly bad, or very bad. E. Teachers
}
\section{Values}{

\itemize{
\item \verb{[1]} very good
\item \verb{[2]} good
\item \verb{[3]} fairly good
\item \verb{[4]} fairly bad
\item \verb{[5]} very bad
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5106/vshow}{GSS Data Explorer page for \code{roletchr}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very good \tab good \tab fairly good \tab fairly bad \tab very bad \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 218 \tab 623 \tab 435 \tab 46 \tab 4 \tab 35 \tab 11 \tab 1372 \cr
   Total \tab 0 \tab 218 \tab 623 \tab 435 \tab 46 \tab 4 \tab 35 \tab 11 \tab 1372 \cr
}
}

\seealso{
Other Children: 
\code{\link{chldeduc}},
\code{\link{chldlove}},
\code{\link{chldmorl}},
\code{\link{chldsafe}},
\code{\link{chldsup}},
\code{\link{roleccp}},
\code{\link{roleclrg}},
\code{\link{rolegp}},
\code{\link{rolepa}}
}
\concept{Children}
\keyword{variable}
