% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_176.R
\name{romance}
\alias{romance}
\title{Do you have a main romantic involvement}
\source{
General Social Survey https://gss.norc.org
}
\description{
romance
}
\details{
Question Do you have a main romantic involvement - a (man/woman) you think of as a steady, a lover, a partner, or whatever?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[3]} has same gender partner
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1649/vshow}{GSS Data Explorer page for \code{romance}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab has same gender partner \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2126 \tab 319 \tab 427 \tab 1 \tab 2 \tab 29 \tab 2904 \cr
   1998 \tab 2109 \tab 272 \tab 415 \tab 3 \tab 3 \tab 30 \tab 2832 \cr
   Total \tab 4235 \tab 591 \tab 842 \tab 4 \tab 5 \tab 59 \tab 5736 \cr
}
}

\seealso{
Other Gender: 
\code{\link{haprom}},
\code{\link{livewith}}
}
\concept{Gender}
\keyword{variable}
