% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_117.R
\name{rotapple}
\alias{rotapple}
\title{R agrees immoral person corrupts society}
\source{
General Social Survey https://gss.norc.org
}
\description{
rotapple
}
\details{
Question Please consider the following statements and tell me whether you agree strongly, agree somewhat, disagree somewhat or disagree strongly with each statement.
C. Immoral actions by one person can corrupt society in general.
}
\section{Values}{

\itemize{
\item \verb{[1]} agree strongly
\item \verb{[2]} agree somewhat
\item \verb{[3]} disagree somewhat
\item \verb{[4]} disagree strongly
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 2006, 2008, 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1083/vshow}{GSS Data Explorer page for \code{rotapple}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab agree strongly \tab agree somewhat \tab disagree somewhat \tab disagree strongly \tab don't know \tab no answer \tab Total \cr
   1988 \tab - \tab 267 \tab 479 \tab 410 \tab 269 \tab 48 \tab 8 \tab 1481 \cr
   2006 \tab 1518 \tab 778 \tab 1063 \tab 712 \tab 373 \tab 61 \tab 5 \tab 4510 \cr
   2008 \tab - \tab 512 \tab 776 \tab 439 \tab 240 \tab 50 \tab 6 \tab 2023 \cr
   2010 \tab - \tab 491 \tab 810 \tab 459 \tab 245 \tab 35 \tab 4 \tab 2044 \cr
   Total \tab 1518 \tab 2048 \tab 3128 \tab 2020 \tab 1127 \tab 194 \tab 23 \tab 10058 \cr
}
}

\seealso{
Other Religion: 
\code{\link{blkwhite}},
\code{\link{decbible}},
\code{\link{decchurh}},
\code{\link{decoths}},
\code{\link{decself}},
\code{\link{gochurch}},
\code{\link{permoral}},
\code{\link{punsin}},
\code{\link{readword}}
}
\concept{Religion}
\keyword{variable}
