% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_77.R
\name{rowngun}
\alias{rowngun}
\title{Does gun belong to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
rowngun
}
\details{
Question Do you happen to have in your home (IF HOUSE:  or garage) any guns or revolvers?
B. Do any of these guns personally belong to you?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} refused
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1980 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/C/- \tab partial \cr
   1989 \tab A/C/- \tab partial \cr
   1990 \tab A/C/- \tab partial \cr
   1991 \tab A/C/- \tab partial \cr
   1993 \tab A/C/- \tab partial \cr
   1994 \tab A/C/- \tab partial \cr
   1996 \tab A/C/- \tab partial \cr
   1998 \tab A/C/- \tab partial \cr
   2000 \tab A/C/- \tab partial \cr
   2002 \tab A/C/- \tab partial \cr
   2004 \tab A/C/- \tab partial \cr
   2006 \tab A/C/- \tab partial \cr
   2008 \tab A/C/- \tab partial \cr
   2010 \tab A/C/- \tab partial \cr
   2012 \tab A/C/- \tab partial \cr
   2014 \tab A/C/- \tab partial \cr
   2016 \tab A/C/- \tab partial \cr
   2018 \tab A/C/- \tab partial \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/C/- \tab partial \cr
   2024 \tab A/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/683/vshow}{GSS Data Explorer page for \code{rowngun}}.
}

\seealso{
Other Core: 
\code{\link{arrest}},
\code{\link{convictd}},
\code{\link{hunt}},
\code{\link{hunt1}},
\code{\link{huntothr}},
\code{\link{lockedup}},
\code{\link{news}()},
\code{\link{ticket}},
\code{\link{tvhours}}
}
\concept{Core}
\keyword{variable}
