% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_445.R
\name{rprtst1}
\alias{rprtst1}
\title{Would you attend a public meeting}
\source{
General Social Survey https://gss.norc.org
}
\description{
rprtst1
}
\details{
Question Would you or would you not do any of the following to protest against a government action you strongly opposed? please CHECK ONE BOX ON EACH LINE.
A. Attend a public meeting organized to protest against the government
}
\section{Values}{

\itemize{
\item \verb{[1]} definietly would
\item \verb{[2]} probably would
\item \verb{[3]} probably would not
\item \verb{[4]} definitely would not
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4090/vshow}{GSS Data Explorer page for \code{rprtst1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definietly would \tab probably would \tab probably would not \tab definitely would not \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1572 \tab 410 \tab 391 \tab 274 \tab 204 \tab 43 \tab 10 \tab 2904 \cr
   Total \tab 1572 \tab 410 \tab 391 \tab 274 \tab 204 \tab 43 \tab 10 \tab 2904 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gendereq}},
\code{\link{grnlaws}},
\code{\link{polint}},
\code{\link{prvdhlth}},
\code{\link{prvdold}},
\code{\link{rprtst15}},
\code{\link{rprtst3}},
\code{\link{rprtst35}},
\code{\link{unionsok}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
