% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_445.R
\name{rprtst3}
\alias{rprtst3}
\title{Would you go on a protest march-demonstration}
\source{
General Social Survey https://gss.norc.org
}
\description{
rprtst3
}
\details{
Question Would you or would you not do any of the following to protest against a government action you strongly opposed? please CHECK ONE BOX ON EACH LINE.
B. Go on a protest march or demonstration
}
\section{Values}{

\itemize{
\item \code{1} definietly would
\item \code{2} probably would
\item \code{3} probably would not
\item \code{4} definitely would not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4091/vshow}{GSS Data Explorer page for \code{rprtst3}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definietly would \tab probably would \tab probably would not \tab definitely would not \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1572 \tab 277 \tab 310 \tab 372 \tab 305 \tab 40 \tab 28 \tab 2904 \cr
   Total \tab 1572 \tab 277 \tab 310 \tab 372 \tab 305 \tab 40 \tab 28 \tab 2904 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gendereq}},
\code{\link{grnlaws}},
\code{\link{polint}},
\code{\link{prvdhlth}},
\code{\link{prvdold}},
\code{\link{rprtst1}},
\code{\link{rprtst15}},
\code{\link{rprtst35}},
\code{\link{unionsok}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
