% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_445.R
\name{rprtst35}
\alias{rprtst35}
\title{How many times gone on a protest march past 5 yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
rprtst35
}
\details{
Question And in the past five years how many times have you done each of the following to protest against a government action you strongly oppose? PLEASE CHECK ONE BOX ON EACH LINE
B. Gone on a protest march or demonstration.
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} once
\item \code{3} more than once
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4093/vshow}{GSS Data Explorer page for \code{rprtst35}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gendereq}},
\code{\link{grnlaws}},
\code{\link{polint}},
\code{\link{prvdhlth}},
\code{\link{prvdold}},
\code{\link{rprtst1}},
\code{\link{rprtst15}},
\code{\link{rprtst3}},
\code{\link{unionsok}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
