% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_568.R
\name{rptcowrk}
\alias{rptcowrk}
\title{Report coworkers threatening secrecy of work}
\source{
General Social Survey https://gss.norc.org
}
\description{
rptcowrk
}
\details{
Question When faced with a conflict between loyalty to one''s employer and loyalty to a co-worker who is violating rules for protecting SECRET and TOP SECRET information, a person should:
}
\section{Values}{

\itemize{
\item \code{1} report coworker
\item \code{2} ask coworker to stop
\item \code{3} report cowrkr if behav cont
\item \code{4} not get involved
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5207/vshow}{GSS Data Explorer page for \code{rptcowrk}}.
}

\seealso{
Other National Security: 
\code{\link{askcrime}},
\code{\link{askdrugs}},
\code{\link{askfinan}},
\code{\link{askforgn}},
\code{\link{askmentl}},
\code{\link{secdocs}},
\code{\link{sectech}}

Other United States: 
\code{\link{askcrime}},
\code{\link{askdrugs}},
\code{\link{askfinan}},
\code{\link{askforgn}},
\code{\link{askmentl}},
\code{\link{secdocs}},
\code{\link{sectech}}
}
\concept{National Security}
\concept{United States}
\keyword{variable}
