% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_644.R
\name{rspgndr}
\alias{rspgndr}
\title{Women vs men responsibility for the home}
\source{
General Social Survey https://gss.norc.org
}
\description{
rspgndr
}
\details{
Question People have different opinions about how much responsibility women and men should have for the home and family. Which of the following statements comes closest to your opinion?
}
\section{Values}{

\itemize{
\item \code{1} women take much more responsibility
\item \code{2} women take somewhat more responsibility
\item \code{3} women and men take equal responsibility
\item \code{4} men take somewhat more responsibility
\item \code{5} men take much more responsibility
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7766/vshow}{GSS Data Explorer page for \code{rspgndr}}.
}

\keyword{variable}
