% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_645.R
\name{rsprltv1}
\alias{rsprltv1}
\title{Offer sibling place to live}
\source{
General Social Survey https://gss.norc.org
}
\description{
rsprltv1
}
\details{
Question People have different opinions about responsibilities towards relatives. Consider a couple living with their two children who has enough space in their home for someone to come and stay. A sibling of one of the couple needs a place to stay for a few mo
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, they definitely should
\item \verb{[7]} no, they definitely should not
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7778/vshow}{GSS Data Explorer page for \code{rsprltv1}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab yes, they definitely should \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab no, they definitely should not \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2386 \tab 285 \tab 137 \tab 178 \tab 246 \tab 67 \tab 27 \tab 80 \tab 120 \tab 9 \tab 9 \tab 3544 \cr
   Total \tab 2386 \tab 285 \tab 137 \tab 178 \tab 246 \tab 67 \tab 27 \tab 80 \tab 120 \tab 9 \tab 9 \tab 3544 \cr
}
}

\keyword{variable}
