% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_645.R
\name{rsprltv2}
\alias{rsprltv2}
\title{Offer parent place to live}
\source{
General Social Survey https://gss.norc.org
}
\description{
rsprltv2
}
\details{
Question Consider another couple living with their two children who has enough space in their home for someone to come and stay. The father of one of them recently became a widower. Although he is in good health, he is feeling lonely. In your opinion, do you think
}
\section{Values}{

\itemize{
\item \code{1} yes, they definitely should
\item \code{7} no, they definitely should not
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7779/vshow}{GSS Data Explorer page for \code{rsprltv2}}.
}

\keyword{variable}
