% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_645.R
\name{rsprltv2}
\alias{rsprltv2}
\title{Offer parent place to live}
\source{
General Social Survey https://gss.norc.org
}
\description{
rsprltv2
}
\details{
Question Consider another couple living with their two children who has enough space in their home for someone to come and stay. The father of one of them recently became a widower. Although he is in good health, he is feeling lonely. In your opinion, do you think
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, they definitely should
\item \verb{[7]} no, they definitely should not
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7779/vshow}{GSS Data Explorer page for \code{rsprltv2}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab yes, they definitely should \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab no, they definitely should not \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2386 \tab 233 \tab 104 \tab 145 \tab 227 \tab 98 \tab 69 \tab 118 \tab 145 \tab 8 \tab 11 \tab 3544 \cr
   Total \tab 2386 \tab 233 \tab 104 \tab 145 \tab 227 \tab 98 \tab 69 \tab 118 \tab 145 \tab 8 \tab 11 \tab 3544 \cr
}
}

\keyword{variable}
