% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_666.R
\name{rstpprnt18}
\alias{rstpprnt18}
\title{R is step-parent of another adult in hh}
\source{
General Social Survey https://gss.norc.org
}
\description{
rstpprnt18
}
\details{
Question To clarify, are you the stepparent of at least one individual 18 or older in your household
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8032/vshow}{GSS Data Explorer page for \code{rstpprnt18}}.
}

\keyword{variable}
