% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_338.R
\name{rundrstd}
\alias{rundrstd}
\title{Did r appear to understand 3-box answer format}
\source{
General Social Survey https://gss.norc.org
}
\description{
rundrstd
}
\details{
Question INTERVIEWER: BY THE TIME YOU HANDED HAND CARD C4 (IMMSTATS) TO THE RESPONDENT, DID HE OR SHE APPEAR TO UNDERSTAND THE 3-BOX ANSWER FORMAT?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3028/vshow}{GSS Data Explorer page for \code{rundrstd}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab other \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2540 \tab 220 \tab 23 \tab 16 \tab 8 \tab 5 \tab 2812 \cr
   Total \tab 2540 \tab 220 \tab 23 \tab 16 \tab 8 \tab 5 \tab 2812 \cr
}
}

\seealso{
Other Immigration: 
\code{\link{bornhome}},
\code{\link{immstats}},
\code{\link{natborn}},
\code{\link{robject}},
\code{\link{travelus}}
}
\concept{Immigration}
\keyword{variable}
