% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_338.R
\name{rundrstd}
\alias{rundrstd}
\title{Did r appear to understand 3-box answer format}
\source{
General Social Survey https://gss.norc.org
}
\description{
rundrstd
}
\details{
Question INTERVIEWER: BY THE TIME YOU HANDED HAND CARD C4 (IMMSTATS) TO THE RESPONDENT, DID HE OR SHE APPEAR TO UNDERSTAND THE 3-BOX ANSWER FORMAT?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3028/vshow}{GSS Data Explorer page for \code{rundrstd}}.
}

\seealso{
Other Immigration: 
\code{\link{bornhome}},
\code{\link{immstats}},
\code{\link{natborn}},
\code{\link{robject}},
\code{\link{travelus}}

Other Interview: 
\code{\link{robject}}
}
\concept{Immigration}
\concept{Interview}
\keyword{variable}
