% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_447.R
\name{runpower}
\alias{runpower}
\title{Electric power should mainly be run by private}
\source{
General Social Survey https://gss.norc.org
}
\description{
runpower
}
\details{
Question Do you think each of the following should mainly be run by private organizations or companies, or by the government? PLEASE CHECK ONE BOX ON EACH LINE.
A. Electric Power
}
\section{Values}{

\itemize{
\item \code{1} by private
\item \code{2} by the government
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4104/vshow}{GSS Data Explorer page for \code{runpower}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{brlawfl}},
\code{\link{brnotax}},
\code{\link{demworks}},
\code{\link{poleff17}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{taxbylaw}},
\code{\link{taxpaid}},
\code{\link{taxspend}}

Other Federal Government: 
\code{\link{demworks}},
\code{\link{poleff17}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{taxspend}}
}
\concept{Federal Government}
\concept{ISSP Role of Gov't}
\keyword{variable}
