% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_86.R
\name{rushed}
\alias{rushed}
\title{How often r feels rushed}
\source{
General Social Survey https://gss.norc.org
}
\description{
rushed
}
\details{
Question In general, how do you feel about your time--would you say you always feel rushed even to do things you have to do, Only sometimes feel rushed, or almost never feel rushed.
}
\section{Values}{

\itemize{
\item \code{1} always
\item \code{2} sometimes
\item \code{3} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1982 \tab -/-/- \tab full \cr
   1996 \tab A/B/C \tab full \cr
   2004 \tab B/C/- \tab partial \cr
   2016 \tab C/-/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/774/vshow}{GSS Data Explorer page for \code{rushed}}.
}

\seealso{
Other Core: 
\code{\link{bored}},
\code{\link{civic}},
\code{\link{feserve}},
\code{\link{imppol}},
\code{\link{meserve}},
\code{\link{pollgood}},
\code{\link{polltrue}},
\code{\link{privacy}},
\code{\link{stresswk}}

Other Time: 
\code{\link{bored}}
}
\concept{Core}
\concept{Gender}
\concept{Time}
\keyword{variable}
