% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_86.R
\name{rushed}
\alias{rushed}
\title{How often r feels rushed}
\source{
General Social Survey https://gss.norc.org
}
\description{
rushed
}
\details{
Question In general, how do you feel about your time--would you say you always feel rushed even to do things you have to do, Only sometimes feel rushed, or almost never feel rushed.
}
\section{Values}{

\itemize{
\item \code{1} always
\item \code{2} sometimes
\item \code{3} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
   B/C/- \tab 2004 \cr
   C/-/- \tab 2016 \cr
   Full \tab 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/774/vshow}{GSS Data Explorer page for \code{rushed}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab always \tab sometimes \tab never \tab no answer \tab don't know \tab Total \cr
   1982 \tab - \tab 442 \tab 967 \tab 443 \tab 8 \tab - \tab 1860 \cr
   1996 \tab 1444 \tab 434 \tab 757 \tab 261 \tab 7 \tab 1 \tab 2904 \cr
   2004 \tab 1828 \tab 304 \tab 513 \tab 160 \tab 6 \tab 1 \tab 2812 \cr
   2016 \tab 1888 \tab 215 \tab 542 \tab 218 \tab 4 \tab - \tab 2867 \cr
   Total \tab 5160 \tab 1395 \tab 2779 \tab 1082 \tab 25 \tab 2 \tab 10443 \cr
}
}

\seealso{
Other Core: 
\code{\link{bored}},
\code{\link{civic}},
\code{\link{feserve}},
\code{\link{imppol}},
\code{\link{meserve}},
\code{\link{pollgood}},
\code{\link{polltrue}},
\code{\link{privacy}},
\code{\link{stresswk}}
}
\concept{Core}
\concept{Gender}
\keyword{variable}
