% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_500.R
\name{rwrkindp}
\alias{rwrkindp}
\title{R can work independently}
\source{
General Social Survey https://gss.norc.org
}
\description{
rwrkindp
}
\details{
Question For each statement about your main job below, please circle one code to show how much you agree or disagree that it applies to your job.
F. I can work independently
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4546/vshow}{GSS Data Explorer page for \code{rwrkindp}}.
}

\seealso{
Other Work Orientation: 
\code{\link{physwrk}},
\code{\link{rflexhrs}},
\code{\link{rhiinc}},
\code{\link{rhlpoths}},
\code{\link{rhlpsoc}},
\code{\link{rintjob}},
\code{\link{rleisure}},
\code{\link{rpromotn}},
\code{\link{xhaustn}}
}
\concept{Work Orientation}
\keyword{variable}
