% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_482.R
\name{rwrknokd}
\alias{rwrknokd}
\title{Did r work aftr marrge but before kids}
\source{
General Social Survey https://gss.norc.org
}
\description{
rwrknokd
}
\details{
Question Did you work outside the home full-time, part-time, or not at all...
A. After marrying and before you had children?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, full-time
\item \verb{[2]} yes,part-time
\item \verb{[3]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4396/vshow}{GSS Data Explorer page for \code{rwrknokd}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes, full-time \tab yes,part-time \tab no \tab don't know \tab no answer \tab Total \cr
   1988 \tab 1148 \tab 192 \tab 38 \tab 82 \tab 7 \tab 14 \tab 1481 \cr
   1994 \tab 1545 \tab 800 \tab 66 \tab 137 \tab 87 \tab 357 \tab 2992 \cr
   2002 \tab 1936 \tab 631 \tab 57 \tab 69 \tab 60 \tab 12 \tab 2765 \cr
   Total \tab 4629 \tab 1623 \tab 161 \tab 288 \tab 154 \tab 383 \tab 7238 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{cohabit}},
\code{\link{divhubby}},
\code{\link{earnsmor}},
\code{\link{evdiv}},
\code{\link{mawork14}},
\code{\link{rwrkbaby}},
\code{\link{rwrkgrwn}},
\code{\link{rwrksch}},
\code{\link{spevdiv}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
