% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_213.R
\name{rxmed}
\alias{rxmed}
\title{X should take prescription medication}
\source{
General Social Survey https://gss.norc.org
}
\description{
rxmed
}
\details{
Question Should NAME do any of the following:
I. Take prescription medication.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2019/vshow}{GSS Data Explorer page for \code{rxmed}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 788 \tab 513 \tab 109 \tab 34 \tab 2904 \cr
   2006 \tab 3073 \tab 1005 \tab 354 \tab 62 \tab 16 \tab 4510 \cr
   2018 \tab 1175 \tab 696 \tab 408 \tab 40 \tab 29 \tab 2348 \cr
   Total \tab 5708 \tab 2489 \tab 1275 \tab 211 \tab 79 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{dofirst}},
\code{\link{meddoc}},
\code{\link{mentldoc}},
\code{\link{mentlhos}},
\code{\link{mentloth}},
\code{\link{otcmed}},
\code{\link{selfhelp}},
\code{\link{sphealer}},
\code{\link{tlkclrgy}}
}
\concept{Mental Health}
\keyword{variable}
