% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_213.R
\name{rxmed}
\alias{rxmed}
\title{X should take prescription medication}
\source{
General Social Survey https://gss.norc.org
}
\description{
rxmed
}
\details{
Question Should NAME do any of the following:
I. Take prescription medication.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2018 \tab B/C/- \tab partial \cr
   2024 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2019/vshow}{GSS Data Explorer page for \code{rxmed}}.
}

\seealso{
Other Mental Health: 
\code{\link{dofirst}},
\code{\link{meddoc}},
\code{\link{mentldoc}},
\code{\link{mentlhos}},
\code{\link{mentloth}},
\code{\link{otcmed}},
\code{\link{selfhelp}},
\code{\link{sphealer}},
\code{\link{tlkclrgy}}

Other Mental Health: 
\code{\link{dofirst}},
\code{\link{meddoc}},
\code{\link{mentldoc}},
\code{\link{mentlhos}},
\code{\link{mentloth}},
\code{\link{otcmed}},
\code{\link{selfhelp}},
\code{\link{sphealer}},
\code{\link{tlkclrgy}}
}
\concept{Mental Health}
\keyword{variable}
