% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_499.R
\name{ryrsfirm}
\alias{ryrsfirm}
\title{Importance of tenure}
\source{
General Social Survey https://gss.norc.org
}
\description{
ryrsfirm
}
\details{
Question Think of two people doing the same kind of work in your place. What Do you personally think is important in deciding how much to pay them?
Looking at the things below, please write ''1'' in the box next to the thing that you think is most important at your place of work. Then write ''2'' next to the thing that you think is next most important, and ''3'' next to the thing that you think is third most important. Leave the other boxes blank.
At my work, deciding on two people doing the same kind of work, how important is...
A. How long the employee has been with the firm
}
\section{Values}{

\itemize{
\item \code{1} most impt
\item \code{2} 2nd most impt
\item \code{3} 3rd most impt
\item \code{4} not chosen
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4533/vshow}{GSS Data Explorer page for \code{ryrsfirm}}.
}

\seealso{
Other Work Orientation: 
\code{\link{hrsmoney}},
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{rsecjob}}

Other Employment: 
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}}

Other Job: 
\code{\link{hrsmoney}},
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}},
\code{\link{rsecjob}}

Other Occupation: 
\code{\link{rageemp}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rpaysame}}
}
\concept{Employment}
\concept{Job}
\concept{Occupation}
\concept{Work Orientation}
\keyword{variable}
