% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_307.R
\name{safehlth}
\alias{safehlth}
\title{Safety and health condition good at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
safehlth
}
\details{
Question Now I''m going to read you a list of statements that might or might not describe your main job. Please tell me whether you strongly agree, agree, disagree, or strongly disagree with each of these statements.
N. The safety and health conditions where I work are good
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2787/vshow}{GSS Data Explorer page for \code{safehlth}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 650 \tab 969 \tab 131 \tab 21 \tab 10 \tab 15 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 667 \tab 923 \tab 89 \tab 17 \tab 24 \tab 14 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 422 \tab 655 \tab 58 \tab 18 \tab 7 \tab 27 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 489 \tab 689 \tab 51 \tab 11 \tab 3 \tab 7 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 473 \tab 863 \tab 70 \tab 7 \tab 4 \tab 31 \tab - \tab 2348 \cr
   2022 \tab 1548 \tab 746 \tab 1045 \tab 135 \tab 26 \tab 12 \tab 13 \tab 19 \tab 3544 \cr
   Total \tab 8338 \tab 3447 \tab 5144 \tab 534 \tab 100 \tab 60 \tab 107 \tab 19 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{partteam}},
\code{\link{prodctiv}},
\code{\link{promteok}},
\code{\link{proudemp}},
\code{\link{setthngs}},
\code{\link{toofewwk}},
\code{\link{trdunion}},
\code{\link{wkdecide}},
\code{\link{wksmooth}}
}
\concept{Quality of Working Life}
\keyword{variable}
