% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_82.R
\name{salsatfn}
\alias{salsatfn}
\title{Salience, financial satisfaction}
\source{
General Social Survey https://gss.norc.org
}
\description{
salsatfn
}
\details{
Question How often would you say that you and your friends think about topics we've been discussing during the interview? would you say that you and your friends think about (READ EACH ITEM A-E) very often, sometimes, or almost never?
C. Satisfaction with their present financial situation
}
\section{Values}{

\itemize{
\item \code{1} very often
\item \code{2} sometimes
\item \code{3} almost never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1978 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/734/vshow}{GSS Data Explorer page for \code{salsatfn}}.
}

\seealso{
Other Core: 
\code{\link{divorce5}},
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{racdif4}},
\code{\link{salabort}},
\code{\link{salfergt}},
\code{\link{sallabor}},
\code{\link{salsci}},
\code{\link{unemp5}}

Other Salience Of Issues: 
\code{\link{salabort}},
\code{\link{salfergt}},
\code{\link{sallabor}},
\code{\link{salsci}}
}
\concept{Core}
\concept{Income}
\concept{Salience Of Issues}
\keyword{variable}
