% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_82.R
\name{salsci}
\alias{salsci}
\title{Salience, scientific community}
\source{
General Social Survey https://gss.norc.org
}
\description{
salsci
}
\details{
Question How often would you say that you and your friends think about topics we've been discussing during the interview? would you say that you and your friends think about (READ EACH ITEM A-E) very often, sometimes, or almost never?
E. The scientific community
}
\section{Values}{

\itemize{
\item \verb{[1]} very often
\item \verb{[2]} sometimes
\item \verb{[3]} almost never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1978 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/736/vshow}{GSS Data Explorer page for \code{salsci}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very often \tab sometimes \tab almost never \tab don't know \tab no answer \tab Total \cr
   1978 \tab - \tab 148 \tab 561 \tab 785 \tab 31 \tab 7 \tab 1532 \cr
   Total \tab 0 \tab 148 \tab 561 \tab 785 \tab 31 \tab 7 \tab 1532 \cr
}
}

\seealso{
Other Core: 
\code{\link{divorce5}},
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{racdif4}},
\code{\link{salabort}},
\code{\link{salfergt}},
\code{\link{sallabor}},
\code{\link{salsatfn}},
\code{\link{unemp5}}
}
\concept{Core}
\keyword{variable}
