% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_383.R
\name{salth2o}
\alias{salth2o}
\title{Ocean water for vegetables}
\source{
General Social Survey https://gss.norc.org
}
\description{
salth2o
}
\details{
Question Ocean water for vegetables.
}
\section{Values}{

\itemize{
\item \code{1} yes, because there is plenty of ocean water
\item \code{2} yes, because ocean water has many natural fertilizers
\item \code{3} no, because ocean water is too salty for plants grown on land
\item \code{4} no, because ocean water is much more polluted than rainwater
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3477/vshow}{GSS Data Explorer page for \code{salth2o}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{anheat}},
\code{\link{anscitst}},
\code{\link{erosion}},
\code{\link{goldfish}},
\code{\link{h2olife}},
\code{\link{lftplane}},
\code{\link{litmstxt}},
\code{\link{maboygrl}},
\code{\link{stormtxt}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
