% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_383.R
\name{salth2o}
\alias{salth2o}
\title{Ocean water for vegetables}
\source{
General Social Survey https://gss.norc.org
}
\description{
salth2o
}
\details{
Question Ocean water for vegetables.
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, because there is plenty of ocean water
\item \verb{[2]} yes, because ocean water has many natural fertilizers
\item \verb{[3]} no, because ocean water is too salty for plants grown on land
\item \verb{[4]} no, because ocean water is much more polluted than rainwater
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3477/vshow}{GSS Data Explorer page for \code{salth2o}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes, because there is plenty of ocean water \tab yes, because ocean water has many natural fertilizers \tab no, because ocean water is too salty for plants grown on land \tab no, because ocean water is much more polluted than rainwater \tab don't know \tab no answer \tab Total \cr
   2008 \tab 518 \tab 23 \tab 36 \tab 1262 \tab 100 \tab 64 \tab 20 \tab 2023 \cr
   Total \tab 518 \tab 23 \tab 36 \tab 1262 \tab 100 \tab 64 \tab 20 \tab 2023 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{anheat}},
\code{\link{anscitst}},
\code{\link{erosion}},
\code{\link{goldfish}},
\code{\link{h2olife}},
\code{\link{lftplane}},
\code{\link{litmstxt}},
\code{\link{maboygrl}},
\code{\link{stormtxt}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
