% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_138.R
\name{samejob}
\alias{samejob}
\title{Current job same or different fromfirst job}
\source{
General Social Survey https://gss.norc.org
}
\description{
samejob
}
\details{
Question When you first came to work for your present employer (IF SELF-EMPLOYED: current organization), was it roughly in the job you have now or was it in a different job?
}
\section{Values}{

\itemize{
\item \verb{[1]} job have now
\item \verb{[2]} different job
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1290/vshow}{GSS Data Explorer page for \code{samejob}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab job have now \tab different job \tab no answer \tab Total \cr
   1991 \tab 605 \tab 638 \tab 267 \tab 7 \tab 1517 \cr
   Total \tab 605 \tab 638 \tab 267 \tab 7 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{jobyears}},
\code{\link{promotng}},
\code{\link{thisjob1}},
\code{\link{thisjob2}},
\code{\link{thisjob3}},
\code{\link{thisjob4}},
\code{\link{thisjob5}},
\code{\link{thisjob6}},
\code{\link{thisjob7}}
}
\concept{Work Organizations}
\keyword{variable}
