% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_608.R
\name{sample}
\alias{sample}
\title{Sampling frame and method}
\source{
General Social Survey https://gss.norc.org
}
\description{
sample
}
\details{
Question The GSS has employed various sampling procedures and frames. Appendix A gives full detail.
}
\section{Values}{

\itemize{
\item \verb{[1]} 1960 sample frame, block quota
\item \verb{[2]} 1970 sample frame, block quota
\item \verb{[3]} 1970 sample frame, full probability
\item \verb{[4]} 1970 sample frame, full probability black oversample
\item \verb{[5]} 1980 black sample frame, full probability black oversample
\item \verb{[6]} 1980 sample frame, full probability
\item \verb{[7]} 1980 sample frame, full probability black oversample
\item \verb{[8]} 1990 sample frame, full probability
\item \verb{[9]} 2000 sample frame, full probability
\item \verb{[10]} 2010 sample frame, full probability
\item \verb{[11]} 2020 address based sample, full probability
\item \verb{[12]} 2022 amerispeak frame, full probability black/hispanic/aanhpi oversample
\item \verb{[13]} 2020 sample frame, full probability
\item \verb{[14]} 2024 amerispeak frame, full probability black/hispanic/aanhpi oversample
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\seealso{
Other Sampling and Weights: 
\code{\link{ballot}},
\code{\link{formwt}},
\code{\link{issp}},
\code{\link{sampcode}},
\code{\link{version}}
}
\concept{Sampling and Weights}
\keyword{variable}
