% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_606.R
\name{saqgene}
\alias{saqgene}
\title{R did gene module for 2004}
\source{
General Social Survey https://gss.norc.org
}
\description{
saqgene
}
\details{
Question C. R did Gene Module
}
\section{Values}{

\itemize{
\item \verb{[1]} r did gene saq
\item \verb{[2]} administered by interviewer
\item \verb{[3]} r did not do gene
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5558/vshow}{GSS Data Explorer page for \code{saqgene}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab r did gene saq \tab administered by interviewer \tab r did not do gene \tab Total \cr
   2004 \tab - \tab 1657 \tab 723 \tab 432 \tab 2812 \cr
   Total \tab 0 \tab 1657 \tab 723 \tab 432 \tab 2812 \cr
}
}

\seealso{
Other Paradata: 
\code{\link{feelevel}},
\code{\link{feeused}},
\code{\link{intid}},
\code{\link{saqissp}},
\code{\link{saqsex}},
\code{\link{whoelse3}},
\code{\link{whoelse4}},
\code{\link{whoelse5}},
\code{\link{whoelse6}}
}
\concept{Paradata}
\keyword{variable}
