% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_606.R
\name{saqsex}
\alias{saqsex}
\title{R did sex module}
\source{
General Social Survey https://gss.norc.org
}
\description{
saqsex
}
\details{
Question B. R did Sex Module
}
\section{Values}{

\itemize{
\item \code{1} r did sex saq
\item \code{2} administered by interviewer
\item \code{3} promised, but did not do
\item \code{4} r did not do sex
\item \code{5} r did, but not sure mode
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 1998, 2000, 2002, 2004, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5557/vshow}{GSS Data Explorer page for \code{saqsex}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab r did sex saq \tab administered by interviewer \tab promised, but did not do \tab r did not do sex \tab r did, but not sure mode \tab Total \cr
   1996 \tab - \tab 1999 \tab 324 \tab 76 \tab 158 \tab 347 \tab 2904 \cr
   1998 \tab - \tab 1867 \tab 301 \tab 94 \tab 235 \tab 335 \tab 2832 \cr
   2000 \tab - \tab 1853 \tab 158 \tab 122 \tab 303 \tab 381 \tab 2817 \cr
   2002 \tab - \tab 1747 \tab 542 \tab - \tab 476 \tab - \tab 2765 \cr
   2004 \tab - \tab 1648 \tab 594 \tab - \tab 570 \tab - \tab 2812 \cr
   2006 \tab 1518 \tab 1948 \tab 455 \tab - \tab 589 \tab - \tab 4510 \cr
   Total \tab 1518 \tab 11062 \tab 2374 \tab 292 \tab 2331 \tab 1063 \tab 18640 \cr
}
}

\seealso{
Other Paradata: 
\code{\link{feelevel}},
\code{\link{feeused}},
\code{\link{intid}},
\code{\link{saqgene}},
\code{\link{saqissp}},
\code{\link{whoelse3}},
\code{\link{whoelse4}},
\code{\link{whoelse5}},
\code{\link{whoelse6}}
}
\concept{Paradata}
\keyword{variable}
