% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_294.R
\name{satdemoc}
\alias{satdemoc}
\title{R satisfied with way democracy works in america}
\source{
General Social Survey https://gss.norc.org
}
\description{
satdemoc
}
\details{
Question On the whole, are you very satisfied, fairly satisfied, not very satisfied, or not at all satisfied with the way democracy works in the United States?
}
\section{Values}{

\itemize{
\item \code{1} very satisfied
\item \code{2} fairly satisfied
\item \code{3} not very satisfied
\item \code{4} not at all satisfied
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2667/vshow}{GSS Data Explorer page for \code{satdemoc}}.
}

\seealso{
Other Freedom: 
\code{\link{choice}},
\code{\link{cntrlife}},
\code{\link{expunpop}},
\code{\link{freeexp1}},
\code{\link{freeexp2}},
\code{\link{freeexp3}},
\code{\link{freemns1}},
\code{\link{freemns2}},
\code{\link{freemns3}},
\code{\link{freenow}},
\code{\link{freeprss}},
\code{\link{howfree}},
\code{\link{inpeace}},
\code{\link{leftlone}},
\code{\link{nogovt}},
\code{\link{partpol}},
\code{\link{rfreenow}},
\code{\link{rhowfree}},
\code{\link{wlthpov}}

Other ANES: 
\code{\link{econstat}},
\code{\link{lkelyvot}},
\code{\link{polintrst}},
\code{\link{ratechall124}},
\code{\link{ratechall1ts}},
\code{\link{ratechall224}},
\code{\link{ratechall2ts}},
\code{\link{ratechall324}},
\code{\link{ratechall3ts}},
\code{\link{ratedemp}},
\code{\link{rateincumb24}},
\code{\link{rateincumbts}},
\code{\link{ratemaga}},
\code{\link{raterepp}},
\code{\link{voteprob}},
\code{\link{whovote24}},
\code{\link{whovote24a}},
\code{\link{whovotets}}
}
\concept{ANES}
\concept{Freedom}
\keyword{variable}
