% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_64.R
\name{satfin}
\alias{satfin}
\title{Satisfaction with financial situation}
\source{
General Social Survey https://gss.norc.org
}
\description{
satfin
}
\details{
Question A. We are interested in how people are getting along financially these days. So far as you and your family are concerned, would you say that you are pretty well satisfied with your present financial situation, more or less satisfied, or not satisfied at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} pretty well satisfied
\item \verb{[2]} more or less satisfied
\item \verb{[3]} not satisfied at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   Full \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/572/vshow}{GSS Data Explorer page for \code{satfin}}.

Counts by year:\tabular{lllllllll}{
   year \tab pretty well satisfied \tab more or less satisfied \tab not satisfied at all \tab don't know \tab no answer \tab iap \tab skipped on web \tab Total \cr
   1972 \tab 522 \tab 720 \tab 366 \tab 1 \tab 4 \tab - \tab - \tab 1613 \cr
   1973 \tab 461 \tab 683 \tab 357 \tab 1 \tab 2 \tab - \tab - \tab 1504 \cr
   1974 \tab 461 \tab 674 \tab 343 \tab 2 \tab 4 \tab - \tab - \tab 1484 \cr
   1975 \tab 458 \tab 627 \tab 394 \tab 4 \tab 7 \tab - \tab - \tab 1490 \cr
   1976 \tab 458 \tab 686 \tab 348 \tab 4 \tab 3 \tab - \tab - \tab 1499 \cr
   1977 \tab 520 \tab 664 \tab 337 \tab 3 \tab 6 \tab - \tab - \tab 1530 \cr
   1978 \tab 518 \tab 646 \tab 365 \tab 2 \tab 1 \tab - \tab - \tab 1532 \cr
   1980 \tab 417 \tab 652 \tab 393 \tab 4 \tab 2 \tab - \tab - \tab 1468 \cr
   1982 \tab 454 \tab 814 \tab 579 \tab 5 \tab 8 \tab - \tab - \tab 1860 \cr
   1983 \tab 456 \tab 657 \tab 479 \tab 5 \tab 2 \tab - \tab - \tab 1599 \cr
   1984 \tab 415 \tab 668 \tab 381 \tab 3 \tab 6 \tab - \tab - \tab 1473 \cr
   1985 \tab 453 \tab 671 \tab 401 \tab 5 \tab 4 \tab - \tab - \tab 1534 \cr
   1986 \tab 445 \tab 626 \tab 395 \tab 3 \tab 1 \tab - \tab - \tab 1470 \cr
   1987 \tab 502 \tab 855 \tab 454 \tab 2 \tab 6 \tab - \tab - \tab 1819 \cr
   1988 \tab 451 \tab 663 \tab 360 \tab 3 \tab 4 \tab - \tab - \tab 1481 \cr
   1989 \tab 469 \tab 678 \tab 385 \tab 4 \tab 1 \tab - \tab - \tab 1537 \cr
   1990 \tab 406 \tab 591 \tab 370 \tab 3 \tab 2 \tab - \tab - \tab 1372 \cr
   1991 \tab 417 \tab 687 \tab 402 \tab 5 \tab 6 \tab - \tab - \tab 1517 \cr
   1993 \tab 430 \tab 712 \tab 454 \tab 7 \tab 3 \tab - \tab - \tab 1606 \cr
   1994 \tab 832 \tab 1365 \tab 779 \tab 5 \tab 11 \tab - \tab - \tab 2992 \cr
   1996 \tab 803 \tab 1287 \tab 806 \tab 4 \tab 4 \tab - \tab - \tab 2904 \cr
   1998 \tab 849 \tab 1255 \tab 720 \tab 5 \tab 3 \tab - \tab - \tab 2832 \cr
   2000 \tab 834 \tab 1261 \tab 708 \tab 9 \tab 5 \tab - \tab - \tab 2817 \cr
   2002 \tab 410 \tab 568 \tab 391 \tab 2 \tab 1 \tab 1393 \tab - \tab 2765 \cr
   2004 \tab 419 \tab 562 \tab 354 \tab 3 \tab 2 \tab 1472 \tab - \tab 2812 \cr
   2006 \tab 896 \tab 1310 \tab 774 \tab 4 \tab 8 \tab 1518 \tab - \tab 4510 \cr
   2008 \tab 573 \tab 819 \tab 624 \tab 5 \tab 2 \tab - \tab - \tab 2023 \cr
   2010 \tab 478 \tab 918 \tab 642 \tab 1 \tab 5 \tab - \tab - \tab 2044 \cr
   2012 \tab 537 \tab 857 \tab 573 \tab 2 \tab 5 \tab - \tab - \tab 1974 \cr
   2014 \tab 698 \tab 1141 \tab 693 \tab 3 \tab 3 \tab - \tab - \tab 2538 \cr
   2016 \tab 816 \tab 1255 \tab 785 \tab 5 \tab 6 \tab - \tab - \tab 2867 \cr
   2018 \tab 737 \tab 1047 \tab 554 \tab 7 \tab 3 \tab - \tab - \tab 2348 \cr
   2021 \tab 1254 \tab 1800 \tab 962 \tab - \tab 1 \tab - \tab 15 \tab 4032 \cr
   2022 \tab 877 \tab 1568 \tab 1081 \tab 7 \tab 4 \tab - \tab 7 \tab 3544 \cr
   Total \tab 19726 \tab 29987 \tab 18009 \tab 128 \tab 135 \tab 4383 \tab 22 \tab 72390 \cr
}
}

\seealso{
Other Core: 
\code{\link{finalter}},
\code{\link{finrela}},
\code{\link{incneed}},
\code{\link{mininc}},
\code{\link{wksub}},
\code{\link{wksub1}},
\code{\link{wksubs}},
\code{\link{wksubs1}},
\code{\link{wksup}}
}
\concept{Core}
\keyword{variable}
