% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_51.R
\name{satfrnd}
\alias{satfrnd}
\title{Friendships}
\source{
General Social Survey https://gss.norc.org
}
\description{
satfrnd
}
\details{
Question For each area of life I am going to name, tell me the number that shows how much satisfaction you get from that area. READ ITEMS A-E. CIRCLE ONE CODE FOR EACH.
D. Your friendships.
}
\section{Values}{

\itemize{
\item \code{1} very great deal
\item \code{2} great deal
\item \code{3} quite a bit
\item \code{4} a fair amount
\item \code{5} some
\item \code{6} a little
\item \code{7} none
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/- \tab full \cr
   1974 \tab -/- \tab full \cr
   1975 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1978 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab B/C \tab partial \cr
   1989 \tab B/C \tab partial \cr
   1990 \tab B/C \tab partial \cr
   1991 \tab B/C \tab partial \cr
   1993 \tab B/C \tab partial \cr
   1994 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/446/vshow}{GSS Data Explorer page for \code{satfrnd}}.
}

\seealso{
Other Core: 
\code{\link{conbus}},
\code{\link{conclerg}},
\code{\link{coneduc}},
\code{\link{confed}},
\code{\link{confinan}},
\code{\link{conlabor}},
\code{\link{conpress}},
\code{\link{satfam}},
\code{\link{sathealt}}

Other Satisfaction: 
\code{\link{satfam}},
\code{\link{sathealt}}

Other Split Ballots: 
\code{\link{conbus}},
\code{\link{conclerg}},
\code{\link{coneduc}},
\code{\link{confed}},
\code{\link{confinan}},
\code{\link{conlabor}},
\code{\link{conpress}},
\code{\link{satfam}},
\code{\link{sathealt}}
}
\concept{Core}
\concept{Satisfaction}
\concept{Split Ballots}
\keyword{variable}
