% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_194.R
\name{sathome}
\alias{sathome}
\title{How satisfied with your purchase}
\source{
General Social Survey https://gss.norc.org
}
\description{
sathome
}
\details{
Question How satisfied are you with your purchase? READ CATEGORIES BELOW.
}
\section{Values}{

\itemize{
\item \code{1} not too satisfied
\item \code{2} pretty satisfied
\item \code{3} extremely satisfied
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1827/vshow}{GSS Data Explorer page for \code{sathome}}.
}

\seealso{
Other Markets: 
\code{\link{lawyer}},
\code{\link{lawyera}},
\code{\link{lawyerb}},
\code{\link{lawyerc}},
\code{\link{realtor1}},
\code{\link{realtore}},
\code{\link{realtorf}},
\code{\link{relagent}},
\code{\link{warrntyh}}

Other Purchases/Transactions: 
\code{\link{lawyer}},
\code{\link{lawyera}},
\code{\link{lawyerb}},
\code{\link{lawyerc}},
\code{\link{realtor1}},
\code{\link{realtore}},
\code{\link{realtorf}},
\code{\link{relagent}},
\code{\link{warrntyh}}
}
\concept{Markets}
\concept{Purchases/Transactions}
\concept{Satisfaction}
\keyword{variable}
