% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_312.R
\name{satjob1}
\alias{satjob1}
\title{Job satisfaction in general}
\source{
General Social Survey https://gss.norc.org
}
\description{
satjob1
}
\details{
Question All in all, how satisfied would you say you are with your job?
}
\section{Values}{

\itemize{
\item \code{1} very satisfied
\item \code{2} somewhat satisfied
\item \code{3} not too satisfied
\item \code{4} not at all satisfied
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2838/vshow}{GSS Data Explorer page for \code{satjob1}}.
}

\seealso{
Other Quality of Working Life: 
\code{\link{arthrtis}},
\code{\link{depress}},
\code{\link{diabetes}},
\code{\link{hyperten}},
\code{\link{knowschd}},
\code{\link{slpprblm}},
\code{\link{stress12}},
\code{\link{usetech}},
\code{\link{weight}}

Other Job: 
\code{\link{knowschd}},
\code{\link{stress12}},
\code{\link{usetech}}
}
\concept{Job}
\concept{Quality of Working Life}
\keyword{variable}
