% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_61.R
\name{satjobhv}
\alias{satjobhv}
\title{Job satisfaction}
\source{
General Social Survey https://gss.norc.org
}
\description{
satjobhv
}
\details{
Question B. On the whole, how satisfied are you with the job you have -- would you say you are very satisfied, moderately satisfied, a little dissatisfied, or very dissatisfied?
}
\section{Values}{

\itemize{
\item \verb{[1]} very satisfied
\item \verb{[2]} moderately satisfied
\item \verb{[3]} a little dissatisfied
\item \verb{[4]} very dissatisfied
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6645/vshow}{GSS Data Explorer page for \code{satjobhv}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very satisfied \tab moderately satisfied \tab a little dissatisfied \tab very dissatisfied \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2561 \tab 118 \tab 134 \tab 32 \tab 10 \tab 10 \tab 2 \tab 2867 \cr
   Total \tab 2561 \tab 118 \tab 134 \tab 32 \tab 10 \tab 10 \tab 2 \tab 2867 \cr
}
}

\keyword{variable}
