% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_195.R
\name{satlawyr}
\alias{satlawyr}
\title{How satisfied with the legal services}
\source{
General Social Survey https://gss.norc.org
}
\description{
satlawyr
}
\details{
Question How satisfied were you with the legal services you received?
}
\section{Values}{

\itemize{
\item \code{1} not too satisfied
\item \code{2} pretty satisfied
\item \code{3} extremely satisfied
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1838/vshow}{GSS Data Explorer page for \code{satlawyr}}.
}

\seealso{
Other Markets: 
\code{\link{fixera}},
\code{\link{fixerb}},
\code{\link{fixhome}},
\code{\link{lawyer1}},
\code{\link{lawyerd}},
\code{\link{lawyere}},
\code{\link{lawyerf}},
\code{\link{lawyerg}},
\code{\link{rellaw}}
}
\concept{Markets}
\keyword{variable}
