% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_466.R
\name{satlife}
\alias{satlife}
\title{R is satisfied with their life}
\source{
General Social Survey https://gss.norc.org
}
\description{
satlife
}
\details{
Question All things considered, how satisfied are you with your life as a whole nowadays?
}
\section{Values}{

\itemize{
\item \code{1} completely satisfied
\item \code{2} mostly satisfied
\item \code{3} somewhat satisfied
\item \code{4} neither satisfied nor dissatisfied
\item \code{5} somewhat dissatisfied
\item \code{6} mostly dissatisfied
\item \code{7} completely dissatisfied
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7275/vshow}{GSS Data Explorer page for \code{satlife}}.
}

\keyword{variable}
