% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_580.R
\name{saturday}
\alias{saturday}
\title{Attended religious services last saturday}
\source{
General Social Survey https://gss.norc.org
}
\description{
saturday
}
\details{
Question On what day or days do you attend religious services during the last seven days? (PROBE, ASK UNTIL "NO" IS GIVEN: "Did you attend religious services on any other days during the last seven days?")
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/C/- \tab partial \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5317/vshow}{GSS Data Explorer page for \code{saturday}}.
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{attreg}},
\code{\link{attrelig}},
\code{\link{attweek}},
\code{\link{cathid}},
\code{\link{charisma}},
\code{\link{eatout}},
\code{\link{friday}},
\code{\link{godoc}},
\code{\link{mediarel}},
\code{\link{monday}},
\code{\link{numdays}},
\code{\link{othrel}},
\code{\link{othrel1}},
\code{\link{othrel2}},
\code{\link{othrel3}},
\code{\link{relid1}},
\code{\link{relid2}},
\code{\link{relid3}},
\code{\link{relidbst}},
\code{\link{religid}},
\code{\link{seefilm}},
\code{\link{sunday}},
\code{\link{thursday}},
\code{\link{tuesday}},
\code{\link{wednesdy}}
}
\concept{Religious Attendance and Identity}
\keyword{variable}
