% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_116.R
\name{saygrace}
\alias{saygrace}
\title{Does r's family say grace before meals}
\source{
General Social Survey https://gss.norc.org
}
\description{
saygrace
}
\details{
Question At your family meals at home, does anyone say grace or give thanks to God aloud before meals?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1079/vshow}{GSS Data Explorer page for \code{saygrace}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1988 \tab - \tab 737 \tab 732 \tab 12 \tab 1481 \cr
   Total \tab 0 \tab 737 \tab 732 \tab 12 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{churhact}},
\code{\link{churhmem}},
\code{\link{god}},
\code{\link{grace16}},
\code{\link{join16}},
\code{\link{lapsed}},
\code{\link{reborn}},
\code{\link{savesoul}},
\code{\link{tvrelig}}
}
\concept{Religion}
\keyword{variable}
