% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_169.R
\name{sbdeg}
\alias{sbdeg}
\title{Highest schl grade of selected r sibling}
\source{
General Social Survey https://gss.norc.org
}
\description{
sbdeg
}
\details{
Question SIBLING'S DEGREE
}
\section{Values}{

\itemize{
\item \code{0} less than high school
\item \code{1} high school
\item \code{2} associate/junior college
\item \code{3} bachelor's
\item \code{4} graduate
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1583/vshow}{GSS Data Explorer page for \code{sbdeg}}.
}

\seealso{
Other Family Mobility: 
\code{\link{sbeduc}},
\code{\link{sbevwork}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpicked}},
\code{\link{sbpres80}},
\code{\link{sbwork1}},
\code{\link{sbwork2}},
\code{\link{sbwrkslf}}

Other Education: 
\code{\link{sbeduc}}

Other Siblings: 
\code{\link{sbeduc}},
\code{\link{sbevwork}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpicked}},
\code{\link{sbpres80}},
\code{\link{sbwork1}},
\code{\link{sbwork2}},
\code{\link{sbwrkslf}}
}
\concept{Education}
\concept{Family Mobility}
\concept{Siblings}
\keyword{variable}
