% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_169.R
\name{sbdeg}
\alias{sbdeg}
\title{Highest schl grade of selected r sibling}
\source{
General Social Survey https://gss.norc.org
}
\description{
sbdeg
}
\details{
Question SIBLING'S DEGREE
}
\section{Values}{

\itemize{
\item \verb{[0]} less than high school
\item \verb{[1]} high school
\item \verb{[2]} associate/junior college
\item \verb{[3]} bachelor's
\item \verb{[4]} graduate
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1583/vshow}{GSS Data Explorer page for \code{sbdeg}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab less than high school \tab high school \tab associate/junior college \tab bachelor's \tab graduate \tab don't know \tab no answer \tab Total \cr
   1994 \tab 378 \tab 390 \tab 1431 \tab 116 \tab 402 \tab 178 \tab 92 \tab 5 \tab 2992 \cr
   Total \tab 378 \tab 390 \tab 1431 \tab 116 \tab 402 \tab 178 \tab 92 \tab 5 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{sbeduc}},
\code{\link{sbevwork}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpicked}},
\code{\link{sbpres80}},
\code{\link{sbwork1}},
\code{\link{sbwork2}},
\code{\link{sbwrkslf}}
}
\concept{Family Mobility}
\keyword{variable}
