% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_169.R
\name{sbevwork}
\alias{sbevwork}
\title{Did r sibling work for 1 or more yr?}
\source{
General Social Survey https://gss.norc.org
}
\description{
sbevwork
}
\details{
Question Has (s/he) ever worked at one job for as long as one year?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1586/vshow}{GSS Data Explorer page for \code{sbevwork}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2210 \tab 528 \tab 135 \tab 64 \tab 55 \tab 2992 \cr
   Total \tab 2210 \tab 528 \tab 135 \tab 64 \tab 55 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{sbdeg}},
\code{\link{sbeduc}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpicked}},
\code{\link{sbpres80}},
\code{\link{sbwork1}},
\code{\link{sbwork2}},
\code{\link{sbwrkslf}}
}
\concept{Family Mobility}
\keyword{variable}
