% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_170.R
\name{sboccindv}
\alias{sboccindv}
\title{Sibling's occ/ind verbatim retrieved}
\source{
General Social Survey https://gss.norc.org
}
\description{
sboccindv
}
\details{
Question Sibling's OCC/IND verbatim recovered
}
\section{Values}{

\itemize{
\item \code{0} no verbatims for sb
\item \code{1} verbatims for sb
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6020/vshow}{GSS Data Explorer page for \code{sboccindv}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab no verbatims for sb \tab verbatims for sb \tab Total \cr
   1994 \tab 54 \tab 555 \tab 2383 \tab 2992 \cr
   Total \tab 54 \tab 555 \tab 2383 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{sbind10}},
\code{\link{sbindstatus}},
\code{\link{sbindtag}},
\code{\link{sbocc10}},
\code{\link{sboccstatus}},
\code{\link{sbocctag}},
\code{\link{sbpres10}},
\code{\link{sbpres105plus}}
}
\concept{Family Mobility}
\keyword{variable}
