% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_170.R
\name{sboccstatus}
\alias{sboccstatus}
\title{Sbocc10 coding status}
\source{
General Social Survey https://gss.norc.org
}
\description{
sboccstatus
}
\details{
Question SBOCC10 Coding Status
}
\section{Values}{

\itemize{
\item \verb{[1]} found, iap
\item \verb{[2]} found, no answer
\item \verb{[3]} found, coded
\item \verb{[4]} not found, iap
\item \verb{[5]} not found, no answer
\item \verb{[6]} not found, coded
\item \verb{[7]} not found, no code assigned
\item \verb{[8]} case not collected in 1994
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6021/vshow}{GSS Data Explorer page for \code{sboccstatus}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab found, iap \tab found, no answer \tab found, coded \tab not found, iap \tab not found, no answer \tab not found, coded \tab not found, no code assigned \tab Total \cr
   1994 \tab - \tab 527 \tab 131 \tab 2280 \tab 17 \tab 1 \tab 26 \tab 10 \tab 2992 \cr
   Total \tab 0 \tab 527 \tab 131 \tab 2280 \tab 17 \tab 1 \tab 26 \tab 10 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{sbind10}},
\code{\link{sbindstatus}},
\code{\link{sbindtag}},
\code{\link{sbocc10}},
\code{\link{sboccindv}},
\code{\link{sbocctag}},
\code{\link{sbpres10}},
\code{\link{sbpres105plus}}
}
\concept{Family Mobility}
\keyword{variable}
