% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_170.R
\name{sbocctag}
\alias{sbocctag}
\title{Sbocc10 coding tag}
\source{
General Social Survey https://gss.norc.org
}
\description{
sbocctag
}
\details{
Question SBOCC10 Coding Tag
}
\section{Values}{

\itemize{
\item \code{1} found
\item \code{2} not found, assigned
\item \code{3} not found, not assigned
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6022/vshow}{GSS Data Explorer page for \code{sbocctag}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab found \tab not found, assigned \tab not found, not assigned \tab Total \cr
   1994 \tab - \tab 2938 \tab 44 \tab 10 \tab 2992 \cr
   Total \tab 0 \tab 2938 \tab 44 \tab 10 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{sbind10}},
\code{\link{sbindstatus}},
\code{\link{sbindtag}},
\code{\link{sbocc10}},
\code{\link{sboccindv}},
\code{\link{sboccstatus}},
\code{\link{sbpres10}},
\code{\link{sbpres105plus}}
}
\concept{Family Mobility}
\keyword{variable}
