% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_169.R
\name{sbpres80}
\alias{sbpres80}
\title{R sibling occupational prestige score (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
sbpres80
}
\details{
Question Respondent's Sibling's Occupation, Prestige, and Industry
A. What kind of work (is/was) (BROTHER'S NAME/SISTER'S NAME) doing? That is, what was (his/her) job called?
OCCUPATION (See REMARKS)
B. IF NOT ALREADY ANSWERED, ASK:  What (does/did) (s/he) actually do in that job? Tell me some of what (are/were) some of (his/her) main duties?
C. What kind of place did (s/he) work for?
INDUSTRY (See REMARKS)
D. IF NOT ALREADY ANSWERED, ASK:  What (do/did) they (make/do)?
E. IF ALREADY ANSWERED, CODE WITHOUT ASKING:  (Is/Was) (he/she) self-employed or (does/did) (he/she) work for someone else?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1588/vshow}{GSS Data Explorer page for \code{sbpres80}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 17 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 59 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 68 \tab 69 \tab 70 \tab 71 \tab 72 \tab 73 \tab 74 \tab 75 \tab 86 \tab dk, na, iap \tab iap \tab Total \cr
   1994 \tab - \tab 10 \tab 4 \tab 15 \tab 7 \tab 50 \tab 38 \tab 56 \tab 21 \tab 5 \tab 21 \tab 67 \tab 68 \tab 90 \tab 51 \tab 73 \tab 34 \tab 59 \tab 49 \tab 111 \tab 11 \tab 19 \tab 62 \tab 82 \tab 11 \tab 84 \tab 39 \tab 68 \tab 32 \tab 113 \tab 99 \tab 19 \tab 71 \tab 25 \tab 168 \tab 39 \tab 18 \tab 21 \tab 3 \tab 7 \tab 27 \tab 10 \tab 19 \tab 49 \tab 30 \tab 6 \tab 7 \tab 74 \tab 42 \tab 58 \tab 10 \tab 22 \tab 2 \tab 8 \tab 9 \tab 13 \tab 33 \tab 14 \tab 20 \tab 719 \tab - \tab 2992 \cr
   Total \tab 0 \tab 10 \tab 4 \tab 15 \tab 7 \tab 50 \tab 38 \tab 56 \tab 21 \tab 5 \tab 21 \tab 67 \tab 68 \tab 90 \tab 51 \tab 73 \tab 34 \tab 59 \tab 49 \tab 111 \tab 11 \tab 19 \tab 62 \tab 82 \tab 11 \tab 84 \tab 39 \tab 68 \tab 32 \tab 113 \tab 99 \tab 19 \tab 71 \tab 25 \tab 168 \tab 39 \tab 18 \tab 21 \tab 3 \tab 7 \tab 27 \tab 10 \tab 19 \tab 49 \tab 30 \tab 6 \tab 7 \tab 74 \tab 42 \tab 58 \tab 10 \tab 22 \tab 2 \tab 8 \tab 9 \tab 13 \tab 33 \tab 14 \tab 20 \tab 719 \tab 0 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{sbdeg}},
\code{\link{sbeduc}},
\code{\link{sbevwork}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpicked}},
\code{\link{sbwork1}},
\code{\link{sbwork2}},
\code{\link{sbwrkslf}}
}
\concept{Family Mobility}
\keyword{variable}
