% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_169.R
\name{sbwrkslf}
\alias{sbwrkslf}
\title{R sibling self-employed?}
\source{
General Social Survey https://gss.norc.org
}
\description{
sbwrkslf
}
\details{
Question Respondent's Sibling's Occupation, Prestige, and Industry
A. What kind of work (is/was) (BROTHER'S NAME/SISTER'S NAME) doing? That is, what was (his/her) job called?
OCCUPATION (See REMARKS)
B. IF NOT ALREADY ANSWERED, ASK:  What (does/did) (s/he) actually do in that job? Tell me some of what (are/were) some of (his/her) main duties?
C. What kind of place did (s/he) work for?
INDUSTRY (See REMARKS)
D. IF NOT ALREADY ANSWERED, ASK:  What (do/did) they (make/do)?
E. IF ALREADY ANSWERED, CODE WITHOUT ASKING:  (Is/Was) (he/she) self-employed or (does/did) (he/she) work for someone else?
}
\section{Values}{

\itemize{
\item \code{1} self employed
\item \code{2} someone else
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1589/vshow}{GSS Data Explorer page for \code{sbwrkslf}}.
}

\seealso{
Other Family Mobility: 
\code{\link{sbdeg}},
\code{\link{sbeduc}},
\code{\link{sbevwork}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpicked}},
\code{\link{sbpres80}},
\code{\link{sbwork1}},
\code{\link{sbwork2}}

Other Occupation: 
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpres80}}

Other Siblings: 
\code{\link{sbdeg}},
\code{\link{sbeduc}},
\code{\link{sbevwork}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpicked}},
\code{\link{sbpres80}},
\code{\link{sbwork1}},
\code{\link{sbwork2}}
}
\concept{Family Mobility}
\concept{Job}
\concept{Occupation}
\concept{Siblings}
\concept{Work}
\keyword{variable}
