% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_169.R
\name{sbwrkslf}
\alias{sbwrkslf}
\title{R sibling self-employed?}
\source{
General Social Survey https://gss.norc.org
}
\description{
sbwrkslf
}
\details{
Question Respondent's Sibling's Occupation, Prestige, and Industry
A. What kind of work (is/was) (BROTHER'S NAME/SISTER'S NAME) doing? That is, what was (his/her) job called?
OCCUPATION (See REMARKS)
B. IF NOT ALREADY ANSWERED, ASK:  What (does/did) (s/he) actually do in that job? Tell me some of what (are/were) some of (his/her) main duties?
C. What kind of place did (s/he) work for?
INDUSTRY (See REMARKS)
D. IF NOT ALREADY ANSWERED, ASK:  What (do/did) they (make/do)?
E. IF ALREADY ANSWERED, CODE WITHOUT ASKING:  (Is/Was) (he/she) self-employed or (does/did) (he/she) work for someone else?
}
\section{Values}{

\itemize{
\item \verb{[1]} self employed
\item \verb{[2]} someone else
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1589/vshow}{GSS Data Explorer page for \code{sbwrkslf}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab self employed \tab someone else \tab don't know \tab no answer \tab Total \cr
   1994 \tab 560 \tab 331 \tab 2039 \tab 19 \tab 43 \tab 2992 \cr
   Total \tab 560 \tab 331 \tab 2039 \tab 19 \tab 43 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{sbdeg}},
\code{\link{sbeduc}},
\code{\link{sbevwork}},
\code{\link{sbind80}},
\code{\link{sbocc80}},
\code{\link{sbpicked}},
\code{\link{sbpres80}},
\code{\link{sbwork1}},
\code{\link{sbwork2}}
}
\concept{Family Mobility}
\keyword{variable}
