% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_112.R
\name{schlhead}
\alias{schlhead}
\title{Did r correctly name local school head}
\source{
General Social Survey https://gss.norc.org
}
\description{
schlhead
}
\details{
Question What is the name of the head of the local school system? (CORRECT ANSWERS INCLUDE THE NAME OF THE HEAD OF THE LOCAL SCHOOL BOARD OR THE NAME OF THE SUPERINTENDENT OF SCHOOLS).
}
\section{Values}{

\itemize{
\item \code{1} correct
\item \code{2} incorrect
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1037/vshow}{GSS Data Explorer page for \code{schlhead}}.
}

\seealso{
Other Socio-political participation: 
\code{\link{fedtrust}},
\code{\link{governor}},
\code{\link{loclobby}},
\code{\link{locself}},
\code{\link{locvote}},
\code{\link{othlobby}},
\code{\link{othself}},
\code{\link{polrally}},
\code{\link{usrep}}
}
\concept{Socio-political participation}
\keyword{variable}
