% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_513.R
\name{schlpray}
\alias{schlpray}
\title{Should there be daily prayer in public schools?}
\source{
General Social Survey https://gss.norc.org
}
\description{
schlpray
}
\details{
Question In your opinion, should there be daily prayers in all public schools?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, definitely
\item \verb{[2]} yes, probably
\item \verb{[3]} no, probably
\item \verb{[4]} no, definitely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4680/vshow}{GSS Data Explorer page for \code{schlpray}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes, definitely \tab yes, probably \tab no, probably \tab no, definitely \tab don't know \tab no answer \tab Total \cr
   1991 \tab 158 \tab 466 \tab 300 \tab 201 \tab 236 \tab 135 \tab 21 \tab 1517 \cr
   Total \tab 158 \tab 466 \tab 300 \tab 201 \tab 236 \tab 135 \tab 21 \tab 1517 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{attend12}},
\code{\link{attendma}},
\code{\link{attendpa}},
\code{\link{densp}},
\code{\link{feelrel}},
\code{\link{prayfreq}},
\code{\link{relactiv}},
\code{\link{relactiv1}}
}
\concept{ISSP Religion}
\keyword{variable}
