% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_145.R
\name{schoolng}
\alias{schoolng}
\title{Importance formal school for you job}
\source{
General Social Survey https://gss.norc.org
}
\description{
schoolng
}
\details{
Question How important were each of the following as sources of the skills that you use in your job?  Would you say very important, important, somewhat important, or not at all important?
A. Formal Schooling.
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{2} important
\item \code{3} somewhat important
\item \code{4} not at all important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1368/vshow}{GSS Data Explorer page for \code{schoolng}}.
}

\seealso{
Other Work Organizations: 
\code{\link{cowrkrel}},
\code{\link{fringe10}},
\code{\link{fringe11}},
\code{\link{fringe7}},
\code{\link{fringe8}},
\code{\link{fringe9}},
\code{\link{jobvshme}},
\code{\link{unmanrel}},
\code{\link{unvote}}

Other Job: 
\code{\link{cowrkrel}},
\code{\link{fringe10}},
\code{\link{fringe11}},
\code{\link{fringe7}},
\code{\link{fringe8}},
\code{\link{fringe9}},
\code{\link{jobvshme}},
\code{\link{unmanrel}}
}
\concept{Education}
\concept{Job}
\concept{Work Organizations}
\keyword{variable}
