% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_266.R
\name{sci30}
\alias{sci30}
\title{R use science site in past 30 days}
\source{
General Social Survey https://gss.norc.org
}
\description{
sci30
}
\details{
Question In the past 30 days, how often have you visited a web site for?
S. Science?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} 1-2 times
\item \code{3} 3-5 times
\item \code{4} more than 5 times
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2412/vshow}{GSS Data Explorer page for \code{sci30}}.
}

\seealso{
Other Information Society: 
\code{\link{buyinf12}},
\code{\link{buyit12}},
\code{\link{cook30}},
\code{\link{hmefin12}},
\code{\link{hobby30}},
\code{\link{invest12}},
\code{\link{person30}},
\code{\link{porn30}},
\code{\link{work12}}

Other Computers And Internet: 
\code{\link{buyinf12}},
\code{\link{buyit12}},
\code{\link{cook30}},
\code{\link{hmefin12}},
\code{\link{hobby30}},
\code{\link{invest12}},
\code{\link{person30}},
\code{\link{porn30}},
\code{\link{work12}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
