% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_368.R
\name{scibnfts}
\alias{scibnfts}
\title{Benefits of sci research outweigh harmful results}
\source{
General Social Survey https://gss.norc.org
}
\description{
scibnfts
}
\details{
Question Now for another type of question. People have frequently noted that scientific research has produced benefits and harmful results. Would you say that, on balance, the benefits of scientific research have outweighed the harmful results, or have the harmful results of scientific research been greater than its benefits?
}
\section{Values}{

\itemize{
\item \verb{[1]} benefits greater
\item \verb{[2]} about equal (if volunteered)
\item \verb{[3]} harmful results greater
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   A/C/- \tab 2010 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3325/vshow}{GSS Data Explorer page for \code{scibnfts}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab benefits greater \tab about equal (if volunteered) \tab harmful results greater \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   2006 \tab 2646 \tab 1284 \tab 328 \tab 129 \tab 122 \tab 1 \tab - \tab 4510 \cr
   2008 \tab 518 \tab 1015 \tab 227 \tab 149 \tab 109 \tab 5 \tab - \tab 2023 \cr
   2010 \tab 1353 \tab 462 \tab 107 \tab 55 \tab 60 \tab 7 \tab - \tab 2044 \cr
   2012 \tab 972 \tab 709 \tab 112 \tab 87 \tab 84 \tab 10 \tab - \tab 1974 \cr
   2014 \tab 1299 \tab 834 \tab 199 \tab 122 \tab 80 \tab 4 \tab - \tab 2538 \cr
   2016 \tab 1477 \tab 1007 \tab 163 \tab 112 \tab 100 \tab 8 \tab - \tab 2867 \cr
   2018 \tab 1173 \tab 861 \tab 124 \tab 117 \tab 64 \tab 9 \tab - \tab 2348 \cr
   Total \tab 9438 \tab 6172 \tab 1260 \tab 771 \tab 619 \tab 44 \tab 0 \tab 18304 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{astrolgy}},
\code{\link{astrosci}},
\code{\link{balneg}},
\code{\link{balpos}},
\code{\link{expdesgn}},
\code{\link{exptext}},
\code{\link{odds1}},
\code{\link{scistudy}},
\code{\link{scitext}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
