% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_387.R
\name{scientbe}
\alias{scientbe}
\title{Scientists want to make life better for avg person}
\source{
General Social Survey https://gss.norc.org
}
\description{
scientbe
}
\details{
Question G. Most scientists want to work on things that will make life better for the average person. Do you strongly agree, agree, disagree, or strongly disagree?
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab B/C/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
   2021 \tab B/C/- \tab partial \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3517/vshow}{GSS Data Explorer page for \code{scientbe}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{accntsci}},
\code{\link{advfront}},
\code{\link{anheat}},
\code{\link{anscitst}},
\code{\link{architct}},
\code{\link{artseals}},
\code{\link{astrolgy}},
\code{\link{astrosci}},
\code{\link{balneg}},
\code{\link{balpos}},
\code{\link{bettrlfe}},
\code{\link{bigbang}},
\code{\link{bigbang1}},
\code{\link{bigbang2}},
\code{\link{biosci}},
\code{\link{bizbstgm}},
\code{\link{bizbstgw}},
\code{\link{bizbstnk}},
\code{\link{bizbsttx}},
\code{\link{bizinfgm}},
\code{\link{bizinfgw}},
\code{\link{bizinfnk}},
\code{\link{bizinftx}},
\code{\link{biznuke}},
\code{\link{boyorgrl}},
\code{\link{caremost}},
\code{\link{chemgen}},
\code{\link{cloning}},
\code{\link{cmprgmng}},
\code{\link{coldeg1}},
\code{\link{colsci}},
\code{\link{colscinm}},
\code{\link{comorsci}},
\code{\link{condrift}},
\code{\link{daynight}},
\code{\link{earthsun}},
\code{\link{eatgm}},
\code{\link{ecoagree}},
\code{\link{ecobsttx}},
\code{\link{ecoinftx}},
\code{\link{econsci}},
\code{\link{electron}},
\code{\link{engagrnk}},
\code{\link{engbrng}},
\code{\link{engbstnk}},
\code{\link{engbtr}},
\code{\link{engda}},
\code{\link{engdgr}},
\code{\link{engdo}},
\code{\link{engearn}},
\code{\link{engfun}},
\code{\link{enggood}},
\code{\link{engint}},
\code{\link{englone}},
\code{\link{engnring}},
\code{\link{engnrsci}},
\code{\link{engnuke}},
\code{\link{engodd}},
\code{\link{engprob}},
\code{\link{engrel}},
\code{\link{engresp}},
\code{\link{engson}},
\code{\link{enhinfnk}},
\code{\link{erosion}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{evolved2}},
\code{\link{expdesgn}},
\code{\link{exptext}},
\code{\link{extinct}},
\code{\link{farming}},
\code{\link{finlcoun}},
\code{\link{fireftng}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{gasregs}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{gmbiz}},
\code{\link{gmmed}},
\code{\link{gmpol}},
\code{\link{goldfish}},
\code{\link{gwbiz}},
\code{\link{gwpol}},
\code{\link{gwsci}},
\code{\link{h2olife}},
\code{\link{histsci}},
\code{\link{hotcore}},
\code{\link{hsbio}},
\code{\link{hschem}},
\code{\link{hsmath}},
\code{\link{hsphys}},
\code{\link{huntbear}},
\code{\link{icecaps}},
\code{\link{icesheet}},
\code{\link{indusgen}},
\code{\link{intecon}},
\code{\link{inteduc}},
\code{\link{intenvir}},
\code{\link{intfarm}},
\code{\link{intintl}},
\code{\link{intmed}},
\code{\link{intmil}},
\code{\link{intrhome}},
\code{\link{intsci}},
\code{\link{intspace}},
\code{\link{inttech}},
\code{\link{inuit}},
\code{\link{inuitway}},
\code{\link{journlsm}},
\code{\link{knwecon}},
\code{\link{knwforgn}},
\code{\link{knwgw}},
\code{\link{knwpolar}},
\code{\link{knwsci}},
\code{\link{lasers}},
\code{\link{lawenfrc}},
\code{\link{leadsci}},
\code{\link{lftplane}},
\code{\link{litmstxt}},
\code{\link{maboygrl}},
\code{\link{majorcol}},
\code{\link{marrcoun}},
\code{\link{medagrgm}},
\code{\link{medagrsc}},
\code{\link{medbstgm}},
\code{\link{medbstsc}},
\code{\link{medinfgm}},
\code{\link{medinfsc}},
\code{\link{medsci}},
\code{\link{medtreat}},
\code{\link{morempg}},
\code{\link{nanoben}},
\code{\link{nanoharm}},
\code{\link{nanoknw1}},
\code{\link{nanoknw2}},
\code{\link{nanotech}},
\code{\link{nanowill}},
\code{\link{newsfrmy}},
\code{\link{newsfrom}},
\code{\link{nextgen}},
\code{\link{noicecap}},
\code{\link{nosun}},
\code{\link{nukeelec}},
\code{\link{nukegen}},
\code{\link{odds1}},
\code{\link{odds2}},
\code{\link{penguins}},
\code{\link{physcsci}},
\code{\link{polaryr1}},
\code{\link{polaryr2}},
\code{\link{polaryr3}},
\code{\link{polaryr4}},
\code{\link{polaryr5}},
\code{\link{polaryr6}},
\code{\link{polaryr7}},
\code{\link{polaryr8}},
\code{\link{polbstgm}},
\code{\link{polbstgw}},
\code{\link{polbstnk}},
\code{\link{polbstsc}},
\code{\link{polbsttx}},
\code{\link{polinfgm}},
\code{\link{polinfgw}},
\code{\link{polinfnk}},
\code{\link{polinfsc}},
\code{\link{polinftx}},
\code{\link{polnuke}},
\code{\link{radioact}},
\code{\link{relbstsc}},
\code{\link{relinfsc}},
\code{\link{salth2o}},
\code{\link{sciagrgw}},
\code{\link{scibnfts}},
\code{\link{scibstgw}},
\code{\link{sciental}},
\code{\link{scientbr}},
\code{\link{scientda}},
\code{\link{scientdn}},
\code{\link{scientdo}},
\code{\link{scientfu}},
\code{\link{scientgo}},
\code{\link{scienthe}},
\code{\link{scientmo}},
\code{\link{scientod}},
\code{\link{scientr}},
\code{\link{scientre}},
\code{\link{scientsn}},
\code{\link{scientwk}},
\code{\link{scifrom}},
\code{\link{scifromy}},
\code{\link{sciimp1}},
\code{\link{sciimp2}},
\code{\link{sciimp3}},
\code{\link{sciimp4}},
\code{\link{sciimp5}},
\code{\link{sciimp6}},
\code{\link{sciimp7}},
\code{\link{sciimp8}},
\code{\link{sciinfgo}},
\code{\link{sciinfgw}},
\code{\link{sciintro}},
\code{\link{scimath}},
\code{\link{scimode}},
\code{\link{scinews1}},
\code{\link{scinews2}},
\code{\link{scinews3}},
\code{\link{scispec}},
\code{\link{scistudy}},
\code{\link{scitext}},
\code{\link{scmed}},
\code{\link{scpol}},
\code{\link{screlig}},
\code{\link{scresrch}},
\code{\link{sealevel}},
\code{\link{seeksci}},
\code{\link{seeksciy}},
\code{\link{seesand}},
\code{\link{slsmnshp}},
\code{\link{socsci}},
\code{\link{solarrev}},
\code{\link{stormtxt}},
\code{\link{tempgen}},
\code{\link{tomatoes}},
\code{\link{toofast}},
\code{\link{tvbears}},
\code{\link{tvinuit}},
\code{\link{tvmeltng}},
\code{\link{tvoil}},
\code{\link{tvozone}},
\code{\link{txbiz}},
\code{\link{txeco}},
\code{\link{txpol}},
\code{\link{upbreath}},
\code{\link{viruses}},
\code{\link{visart}},
\code{\link{vislib}},
\code{\link{visnhist}},
\code{\link{vissci}},
\code{\link{viszoo}},
\code{\link{watergen}},
\code{\link{weighing}},
\code{\link{whichsci}}

Other GSS Next: 
\code{\link{abbelief}},
\code{\link{abgender}},
\code{\link{abhelp1}},
\code{\link{abhelp2}},
\code{\link{abhelp3}},
\code{\link{abhelp4}},
\code{\link{abmoral}},
\code{\link{advfront}},
\code{\link{biokids}},
\code{\link{blkdsrv}},
\code{\link{blkmblty}},
\code{\link{blktry}},
\code{\link{brv16}},
\code{\link{brv16bronum}},
\code{\link{brv16cuznum}},
\code{\link{brv16dadnum}},
\code{\link{brv16filnum}},
\code{\link{brv16frndnum}},
\code{\link{brv16gmanum}},
\code{\link{brv16gpanum}},
\code{\link{brv16grand}},
\code{\link{brv16milnum}},
\code{\link{brv16momnum}},
\code{\link{brv16oth}},
\code{\link{brv16othnum}},
\code{\link{brv16par}},
\code{\link{brv16partnum}},
\code{\link{brv16sib}},
\code{\link{brv16silnum}},
\code{\link{brv16sisnum}},
\code{\link{brv16sp}},
\code{\link{brv16spnum}},
\code{\link{brv5}},
\code{\link{brv5bronum}},
\code{\link{brv5child}},
\code{\link{brv5chinum}},
\code{\link{brv5cowknum}},
\code{\link{brv5cuznum}},
\code{\link{brv5dadnum}},
\code{\link{brv5daunum}},
\code{\link{brv5filnum}},
\code{\link{brv5frndnum}},
\code{\link{brv5gmanum}},
\code{\link{brv5gpanum}},
\code{\link{brv5grand}},
\code{\link{brv5milnum}},
\code{\link{brv5momnum}},
\code{\link{brv5oth}},
\code{\link{brv5othnum}},
\code{\link{brv5par}},
\code{\link{brv5partnum}},
\code{\link{brv5sib}},
\code{\link{brv5silnum}},
\code{\link{brv5sisnum}},
\code{\link{brv5sonnum}},
\code{\link{brv5sp}},
\code{\link{brv5spnum}},
\code{\link{confedvacy}},
\code{\link{conpharvacy}},
\code{\link{covemply}},
\code{\link{covid12}},
\code{\link{femself}},
\code{\link{firstkidsex}},
\code{\link{fluvax}},
\code{\link{fund_next}},
\code{\link{gestate}},
\code{\link{malekids}},
\code{\link{mascself}},
\code{\link{nexteligible}},
\code{\link{nextgen}},
\code{\link{nonbinkids}},
\code{\link{pandinc}},
\code{\link{pandmet}},
\code{\link{scientgo}},
\code{\link{scienthe}},
\code{\link{toofast}},
\code{\link{vaxhstncy}},
\code{\link{vaxkids}},
\code{\link{vaxsafe}}
}
\concept{GSS Next}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
