% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_387.R
\name{scientbe}
\alias{scientbe}
\title{Scientists want to make life better for avg person}
\source{
General Social Survey https://gss.norc.org
}
\description{
scientbe
}
\details{
Question G. Most scientists want to work on things that will make life better for the average person. Do you strongly agree, agree, disagree, or strongly disagree?
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab B/C/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
   2021 \tab B/C/- \tab partial \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3517/vshow}{GSS Data Explorer page for \code{scientbe}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{sciental}},
\code{\link{scientdn}},
\code{\link{scientdo}},
\code{\link{scientfu}},
\code{\link{scientgo}},
\code{\link{scienthe}},
\code{\link{scientod}},
\code{\link{scientr}},
\code{\link{scientre}}

Other GSS Next: 
\code{\link{scientgo}},
\code{\link{scienthe}}
}
\concept{GSS Next}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
