% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_387.R
\name{scientod}
\alias{scientod}
\title{Scientists odd and peculiar}
\source{
General Social Survey https://gss.norc.org
}
\description{
scientod
}
\details{
Question F. Scientists are apt to be odd and peculiar people. Do you strongly agree, agree, disagree, or strongly disagree?
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2016, 2018 \cr
   B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3516/vshow}{GSS Data Explorer page for \code{scientod}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2012 \tab 1457 \tab 26 \tab 144 \tab 269 \tab 30 \tab 43 \tab 5 \tab 1974 \cr
   2016 \tab 1477 \tab 132 \tab 584 \tab 520 \tab 88 \tab 51 \tab 15 \tab 2867 \cr
   2018 \tab 1173 \tab 115 \tab 465 \tab 465 \tab 84 \tab 43 \tab 3 \tab 2348 \cr
   Total \tab 4107 \tab 273 \tab 1193 \tab 1254 \tab 202 \tab 137 \tab 23 \tab 7189 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{sciental}},
\code{\link{scientbe}},
\code{\link{scientdn}},
\code{\link{scientdo}},
\code{\link{scientfu}},
\code{\link{scientgo}},
\code{\link{scienthe}},
\code{\link{scientr}},
\code{\link{scientre}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
