% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_387.R
\name{scientr}
\alias{scientr}
\title{Consider career in science}
\source{
General Social Survey https://gss.norc.org
}
\description{
scientr
}
\details{
Question C. Have you ever considered working in a science-related career?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3509/vshow}{GSS Data Explorer page for \code{scientr}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{sciental}},
\code{\link{scientbe}},
\code{\link{scientdn}},
\code{\link{scientdo}},
\code{\link{scientfu}},
\code{\link{scientgo}},
\code{\link{scienthe}},
\code{\link{scientod}},
\code{\link{scientre}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
