% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_386.R
\name{scientsn}
\alias{scientsn}
\title{Happy if son scientist}
\source{
General Social Survey https://gss.norc.org
}
\description{
scientsn
}
\details{
Question B. If you had a son, how would you feel if he wanted to be a scientist"would you feel happy, unhappy, or would you not care one way or the other?
}
\section{Values}{

\itemize{
\item \verb{[1]} happy
\item \verb{[2]} not care
\item \verb{[3]} unhappy
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3508/vshow}{GSS Data Explorer page for \code{scientsn}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab happy \tab not care \tab unhappy \tab don't know \tab no answer \tab Total \cr
   2012 \tab 1457 \tab 426 \tab 81 \tab 3 \tab 3 \tab 4 \tab 1974 \cr
   Total \tab 1457 \tab 426 \tab 81 \tab 3 \tab 3 \tab 4 \tab 1974 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{cloning}},
\code{\link{newsfrmy}},
\code{\link{scientda}},
\code{\link{scifromy}},
\code{\link{sciinfgo}},
\code{\link{scinews1}},
\code{\link{scinews2}},
\code{\link{scinews3}},
\code{\link{seeksciy}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
