% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_519.R
\name{scifaith}
\alias{scifaith}
\title{Believe too much in science, not enough faith}
\source{
General Social Survey https://gss.norc.org
}
\description{
scifaith
}
\details{
Question How much do you agree or disagree with each of these statements?
A. We believe too often in science, and not enough in feelings and faith.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2010 \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4721/vshow}{GSS Data Explorer page for \code{scifaith}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1993 \tab 49 \tab 201 \tab 667 \tab 329 \tab 218 \tab 57 \tab 33 \tab 52 \tab 1606 \cr
   1994 \tab 1606 \tab 161 \tab 558 \tab 313 \tab 233 \tab 41 \tab 41 \tab 39 \tab 2992 \cr
   2000 \tab 1541 \tab 163 \tab 478 \tab 294 \tab 188 \tab 61 \tab 24 \tab 68 \tab 2817 \cr
   2010 \tab 614 \tab 128 \tab 468 \tab 359 \tab 314 \tab 106 \tab 50 \tab 5 \tab 2044 \cr
   Total \tab 3810 \tab 653 \tab 2171 \tab 1295 \tab 953 \tab 265 \tab 148 \tab 164 \tab 9459 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{anrights}},
\code{\link{grnecon}},
\code{\link{harmgood}},
\code{\link{harmsgrn}},
\code{\link{postmat1}},
\code{\link{postmat2}},
\code{\link{privent}},
\code{\link{scigrn}},
\code{\link{sciworse}}
}
\concept{ISSP Environment}
\keyword{variable}
