% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_376.R
\name{sciimp8}
\alias{sciimp8}
\title{How important: research takes place in a laboratory}
\source{
General Social Survey https://gss.norc.org
}
\description{
sciimp8
}
\details{
Question Now I’m going to read you some statements about science and scientists. Please look at Card B19. How important are each of the following in making something scientific?
H. The research takes place in a laboratory. (Would you say this is very important, pretty important, not too important, or not important at all in making something scientific?)
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{2} pretty important
\item \code{3} not too important
\item \code{4} not important at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3405/vshow}{GSS Data Explorer page for \code{sciimp8}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very important \tab pretty important \tab not too important \tab not important at all \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 761 \tab 579 \tab 359 \tab 88 \tab 70 \tab 7 \tab 4510 \cr
   Total \tab 2646 \tab 761 \tab 579 \tab 359 \tab 88 \tab 70 \tab 7 \tab 4510 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{histsci}},
\code{\link{physcsci}},
\code{\link{sciimp2}},
\code{\link{sciimp3}},
\code{\link{sciimp4}},
\code{\link{sciimp5}},
\code{\link{sciimp6}},
\code{\link{sciimp7}},
\code{\link{socsci}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
