% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_381.R
\name{scimode}
\alias{scimode}
\title{Was saq self administered or interview administered}
\source{
General Social Survey https://gss.norc.org
}
\description{
scimode
}
\details{
Question Is the SAQ self administered or interview administered?
}
\section{Values}{

\itemize{
\item \code{1} assigned and done as interviewed
\item \code{2} assigned and done as saq
\item \code{3} assigned as saq, done in interview due to r's request/inability to do as saq
\item \code{4} assigned as saq, done as interview on phone
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3451/vshow}{GSS Data Explorer page for \code{scimode}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{intecon}},
\code{\link{inteduc}},
\code{\link{intfarm}},
\code{\link{intintl}},
\code{\link{intmed}},
\code{\link{intsci}},
\code{\link{inttech}},
\code{\link{nanoben}},
\code{\link{nanoharm}}
}
\concept{Interview}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
