% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_386.R
\name{scinews2}
\alias{scinews2}
\title{Magazine printed or online}
\source{
General Social Survey https://gss.norc.org
}
\description{
scinews2
}
\details{
Question You said you get most of your information about science and technology from magazines. would that be printed magazines or online magazines?
}
\section{Values}{

\itemize{
\item \code{1} printed magazines
\item \code{2} online magazines
\item \code{3} other (specify)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab C/- \tab partial \cr
   2012 \tab B/C \tab partial \cr
   2014 \tab B/C \tab partial \cr
   2016 \tab A/B \tab partial \cr
   2018 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3501/vshow}{GSS Data Explorer page for \code{scinews2}}.
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{cloning}},
\code{\link{newsfrmy}},
\code{\link{scientda}},
\code{\link{scientsn}},
\code{\link{scifromy}},
\code{\link{sciinfgo}},
\code{\link{scinews1}},
\code{\link{scinews3}},
\code{\link{seeksciy}}

Other Science: 
\code{\link{newsfrmy}},
\code{\link{scientda}},
\code{\link{scientsn}},
\code{\link{scifromy}},
\code{\link{sciinfgo}},
\code{\link{scinews1}},
\code{\link{scinews3}},
\code{\link{seeksciy}}
}
\concept{Science}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
