% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_386.R
\name{scinews3}
\alias{scinews3}
\title{Where online get info}
\source{
General Social Survey https://gss.norc.org
}
\description{
scinews3
}
\details{
Question You said you get most of your information about science and technology from the Internet. What is the place you are most likely to go on the Internet for science and technology information—online newspapers, online magazines, or some other place on the Internet?
you most likely to go on the Internet for science and technology information?
}
\section{Values}{

\itemize{
\item \code{1} online newspapers
\item \code{2} online magazines
\item \code{3} science site
\item \code{4} news site
\item \code{5} electronic books & reports
\item \code{6} wikipedia
\item \code{7} government site
\item \code{8} social media (facebook, youtube, twitter)
\item \code{10} other
\item \code{11} search engine (google, bing)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2016, 2018 \cr
   B/C \tab 2012, 2014 \cr
   C/- \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3502/vshow}{GSS Data Explorer page for \code{scinews3}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab online newspapers \tab online magazines \tab science site \tab don't know \tab no answer \tab news site \tab electronic books & reports \tab wikipedia \tab government site \tab social media (facebook, youtube, twitter) \tab other \tab search engine (google, bing) \tab Total \cr
   2010 \tab 1954 \tab 51 \tab 2 \tab 17 \tab 1 \tab 19 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1580 \tab 90 \tab 48 \tab 33 \tab 11 \tab 5 \tab 21 \tab 2 \tab 2 \tab 4 \tab 13 \tab 26 \tab 139 \tab 1974 \cr
   2014 \tab 2012 \tab 130 \tab 74 \tab 44 \tab 14 \tab 4 \tab 35 \tab - \tab 6 \tab 2 \tab - \tab 39 \tab 178 \tab 2538 \cr
   2016 \tab 2146 \tab 177 \tab 103 \tab 42 \tab 16 \tab 7 \tab 45 \tab - \tab 7 \tab 4 \tab 27 \tab 36 \tab 257 \tab 2867 \cr
   2018 \tab 1698 \tab 161 \tab 85 \tab 51 \tab 11 \tab 2 \tab 24 \tab 3 \tab 7 \tab 11 \tab 34 \tab 41 \tab 220 \tab 2348 \cr
   Total \tab 9390 \tab 609 \tab 312 \tab 187 \tab 53 \tab 37 \tab 125 \tab 5 \tab 22 \tab 21 \tab 74 \tab 142 \tab 794 \tab 11771 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{cloning}},
\code{\link{newsfrmy}},
\code{\link{scientda}},
\code{\link{scientsn}},
\code{\link{scifromy}},
\code{\link{sciinfgo}},
\code{\link{scinews1}},
\code{\link{scinews2}},
\code{\link{seeksciy}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
