% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_113.R
\name{scipry}
\alias{scipry}
\title{Scientists pry into things they oughtn't}
\source{
General Social Survey https://gss.norc.org
}
\description{
scipry
}
\details{
Question Here are some things that have been said about science. Would you tell me if you tend to agree or disagree with them?
C. Scientists always seem to be prying into things that they really ought to stay out of.
}
\section{Values}{

\itemize{
\item \code{1} agree
\item \code{2} disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1047/vshow}{GSS Data Explorer page for \code{scipry}}.
}

\seealso{
Other Religion: 
\code{\link{scichng}},
\code{\link{scimoral}},
\code{\link{scisolve}},
\code{\link{switched}}
}
\concept{Religion}
\keyword{variable}
