% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_367.R
\name{scispec}
\alias{scispec}
\title{Science is too concerned with theory and speculation}
\source{
General Social Survey https://gss.norc.org
}
\description{
scispec
}
\details{
Question I'm going to read to you some statements like those you might find in a newspaper or magazine article. For each statement, please tell me if you strongly agree, agree, disagree, or strongly disagree.
E. Science is too concerned with theory and speculation to be of much use in making concrete government policy decisions that will affect the way we live.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3320/vshow}{GSS Data Explorer page for \code{scispec}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 100 \tab 561 \tab 852 \tab 211 \tab 135 \tab 5 \tab 4510 \cr
   Total \tab 2646 \tab 100 \tab 561 \tab 852 \tab 211 \tab 135 \tab 5 \tab 4510 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{advfront}},
\code{\link{bettrlfe}},
\code{\link{leadsci}},
\code{\link{newsfrom}},
\code{\link{nextgen}},
\code{\link{scifrom}},
\code{\link{seeksci}},
\code{\link{toofast}},
\code{\link{whichsci}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
