% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_521.R
\name{scitest4}
\alias{scitest4}
\title{Humans evolved from animals}
\source{
General Social Survey https://gss.norc.org
}
\description{
scitest4
}
\details{
Question For each statement below, just check the box that comes closest to your opinion of how true it is. In your opinion, how true is this?
D. Human beings developed from earlier species of animals.
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely true
\item \verb{[2]} probably true
\item \verb{[3]} probably not true
\item \verb{[4]} definitely not true
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4744/vshow}{GSS Data Explorer page for \code{scitest4}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely true \tab probably true \tab probably not true \tab definitely not true \tab don't know \tab no answer \tab Total \cr
   1993 \tab 49 \tab 205 \tab 445 \tab 200 \tab 500 \tab 120 \tab 87 \tab 1606 \cr
   1994 \tab 1606 \tab 188 \tab 406 \tab 224 \tab 410 \tab 113 \tab 45 \tab 2992 \cr
   2000 \tab 1541 \tab 172 \tab 334 \tab 194 \tab 395 \tab 115 \tab 66 \tab 2817 \cr
   Total \tab 3196 \tab 565 \tab 1185 \tab 618 \tab 1305 \tab 348 \tab 198 \tab 7415 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{grnsol}},
\code{\link{grntest1}},
\code{\link{grntest2}},
\code{\link{ihlpgrn}},
\code{\link{scitest1}},
\code{\link{scitest2}},
\code{\link{scitest3}},
\code{\link{scitest5}},
\code{\link{toodifme}}
}
\concept{ISSP Environment}
\keyword{variable}
