% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_521.R
\name{scitest4}
\alias{scitest4}
\title{Humans evolved from animals}
\source{
General Social Survey https://gss.norc.org
}
\description{
scitest4
}
\details{
Question For each statement below, just check the box that comes closest to your opinion of how true it is. In your opinion, how true is this?
D. Human beings developed from earlier species of animals.
}
\section{Values}{

\itemize{
\item \code{1} definitely true
\item \code{2} probably true
\item \code{3} probably not true
\item \code{4} definitely not true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4744/vshow}{GSS Data Explorer page for \code{scitest4}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{grnsol}},
\code{\link{grntest1}},
\code{\link{grntest2}},
\code{\link{ihlpgrn}},
\code{\link{scitest1}},
\code{\link{scitest2}},
\code{\link{scitest3}},
\code{\link{scitest5}},
\code{\link{toodifme}}

Other Knowledge: 
\code{\link{grntest1}},
\code{\link{grntest2}},
\code{\link{scitest1}},
\code{\link{scitest2}},
\code{\link{scitest3}},
\code{\link{scitest5}}

Other Science: 
\code{\link{scitest1}},
\code{\link{scitest2}},
\code{\link{scitest3}},
\code{\link{scitest5}}
}
\concept{ISSP Environment}
\concept{Knowledge}
\concept{Science}
\keyword{variable}
