% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_544.R
\name{scitesty}
\alias{scitesty}
\title{Human beings developed from earlier species of animal}
\source{
General Social Survey https://gss.norc.org
}
\description{
scitesty
}
\details{
Question From the statement I’ll read, look at this card and tell me which answer comes closest to your opinion on how true it is.
Human beings developed from earlier species of animals.
}
\section{Values}{

\itemize{
\item \code{1} definitely true
\item \code{2} probably true
\item \code{3} probably not true
\item \code{4} definitely not true
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4973/vshow}{GSS Data Explorer page for \code{scitesty}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely true \tab probably true \tab probably not true \tab definitely not true \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 219 \tab 425 \tab 218 \tab 580 \tab 22 \tab 8 \tab 2812 \cr
   Total \tab 1340 \tab 219 \tab 425 \tab 218 \tab 580 \tab 22 \tab 8 \tab 2812 \cr
}
}

\seealso{
Other Misc II: 
\code{\link{creation}}
}
\concept{Misc II}
\keyword{variable}
