% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_519.R
\name{sciworse}
\alias{sciworse}
\title{Humans change nature for the worse}
\source{
General Social Survey https://gss.norc.org
}
\description{
sciworse
}
\details{
Question How much do you agree or disagree with each of these statements?
C. Any change humans cause in nature - no matter how scientific - is likely to make things worse.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4723/vshow}{GSS Data Explorer page for \code{sciworse}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{anrights}},
\code{\link{grnecon}},
\code{\link{harmgood}},
\code{\link{harmsgrn}},
\code{\link{postmat1}},
\code{\link{postmat2}},
\code{\link{privent}},
\code{\link{scifaith}},
\code{\link{scigrn}}
}
\concept{ISSP Environment}
\keyword{variable}
