% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_380.R
\name{sealevel}
\alias{sealevel}
\title{Effect of global warming: sea level flood coastal areas}
\source{
General Social Survey https://gss.norc.org
}
\description{
sealevel
}
\details{
Question Please look at Card B24. Scientists predict that global warming may soon have big effects on the polar regions. I will describe some of these possible effects and, for each one, please say whether it would bother you a great deal, some, a little, or not at all if it actually happened.
B. Sea level may rise by more than 20 feet, flooding coastal areas.
}
\section{Values}{

\itemize{
\item \verb{[1]} a great deal
\item \verb{[2]} some
\item \verb{[3]} a little
\item \verb{[4]} not at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
   B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3439/vshow}{GSS Data Explorer page for \code{sealevel}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab a great deal \tab some \tab a little \tab not at all \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2646 \tab 1256 \tab 350 \tab 124 \tab 82 \tab 50 \tab 2 \tab 4510 \cr
   2010 \tab 1338 \tab 461 \tab 114 \tab 64 \tab 46 \tab 11 \tab 10 \tab 2044 \cr
   Total \tab 3984 \tab 1717 \tab 464 \tab 188 \tab 128 \tab 61 \tab 12 \tab 6554 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{artseals}},
\code{\link{caremost}},
\code{\link{inuitway}},
\code{\link{nanoknw1}},
\code{\link{nanoknw2}},
\code{\link{nanotech}},
\code{\link{nanowill}},
\code{\link{noicecap}},
\code{\link{penguins}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
