% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_569.R
\name{secbudgt}
\alias{secbudgt}
\title{Government maintain secrecy on the intelligence budget}
\source{
General Social Survey https://gss.norc.org
}
\description{
secbudgt
}
\details{
Question Should or should not the government maintain a high level of secrecy surrounding: (READ A-D)
D. The US intelligence budget
}
\section{Values}{

\itemize{
\item \code{1} definitely should
\item \code{2} probably should
\item \code{3} probably should not
\item \code{4} definitely should not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5222/vshow}{GSS Data Explorer page for \code{secbudgt}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should \tab probably should \tab probably should not \tab definitely should not \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 382 \tab 397 \tab 351 \tab 241 \tab 87 \tab 2 \tab 2904 \cr
   1998 \tab 1445 \tab 320 \tab 453 \tab 318 \tab 197 \tab 87 \tab 12 \tab 2832 \cr
   Total \tab 2889 \tab 702 \tab 850 \tab 669 \tab 438 \tab 174 \tab 14 \tab 5736 \cr
}
}

\seealso{
Other National Security: 
\code{\link{askcomp}},
\code{\link{askdrink}},
\code{\link{askfrbiz}},
\code{\link{askfrtrv}},
\code{\link{asksexor}},
\code{\link{secdiplo}},
\code{\link{secmilop}},
\code{\link{secprvcy}},
\code{\link{secterr}}
}
\concept{National Security}
\keyword{variable}
