% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_569.R
\name{secdiplo}
\alias{secdiplo}
\title{Government maintain secrecy on diplomatic initiatives}
\source{
General Social Survey https://gss.norc.org
}
\description{
secdiplo
}
\details{
Question Should or should not the government maintain a high level of secrecy surrounding: (READ A-D)
A. Diplomatic initiatives
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely should
\item \verb{[2]} probably should
\item \verb{[3]} probably should not
\item \verb{[4]} definitely should not
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5219/vshow}{GSS Data Explorer page for \code{secdiplo}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should \tab probably should \tab probably should not \tab definitely should not \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 491 \tab 573 \tab 180 \tab 68 \tab 146 \tab 2 \tab 2904 \cr
   1998 \tab 1445 \tab 400 \tab 624 \tab 145 \tab 65 \tab 142 \tab 11 \tab 2832 \cr
   Total \tab 2889 \tab 891 \tab 1197 \tab 325 \tab 133 \tab 288 \tab 13 \tab 5736 \cr
}
}

\seealso{
Other National Security: 
\code{\link{askcomp}},
\code{\link{askdrink}},
\code{\link{askfrbiz}},
\code{\link{askfrtrv}},
\code{\link{asksexor}},
\code{\link{secbudgt}},
\code{\link{secmilop}},
\code{\link{secprvcy}},
\code{\link{secterr}}
}
\concept{National Security}
\keyword{variable}
