% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_568.R
\name{secdocs}
\alias{secdocs}
\title{Does govt classify too many documents as secrets}
\source{
General Social Survey https://gss.norc.org
}
\description{
secdocs
}
\details{
Question Given the world situation, the government protects too many documents by classifying them as SECRET and TOP SECRET. Do you . . .
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5206/vshow}{GSS Data Explorer page for \code{secdocs}}.
}

\seealso{
Other National Security: 
\code{\link{askcrime}},
\code{\link{askdrugs}},
\code{\link{askfinan}},
\code{\link{askforgn}},
\code{\link{askmentl}},
\code{\link{rptcowrk}},
\code{\link{sectech}}

Other United States: 
\code{\link{askcrime}},
\code{\link{askdrugs}},
\code{\link{askfinan}},
\code{\link{askforgn}},
\code{\link{askmentl}},
\code{\link{rptcowrk}},
\code{\link{sectech}}
}
\concept{National Security}
\concept{United States}
\keyword{variable}
