% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_305.R
\name{secondwk}
\alias{secondwk}
\title{R has job other than main}
\source{
General Social Survey https://gss.norc.org
}
\description{
secondwk
}
\details{
Question Do you have any jobs besides your main job or do any other work for pay?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2773/vshow}{GSS Data Explorer page for \code{secondwk}}.
}

\seealso{
Other Quality of Working Life: 
\code{\link{famvswk}},
\code{\link{famwkoff}},
\code{\link{hrsrelax}},
\code{\link{learnnew}},
\code{\link{whywkhme}},
\code{\link{wkvsfam}},
\code{\link{workdiff}},
\code{\link{workfast}},
\code{\link{wrkhome}}

Other Job: 
\code{\link{famwkoff}},
\code{\link{learnnew}},
\code{\link{workdiff}},
\code{\link{workfast}}
}
\concept{Job}
\concept{Quality of Working Life}
\keyword{variable}
