% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_303.R
\name{secopin}
\alias{secopin}
\title{Would r seek second opinion}
\source{
General Social Survey https://gss.norc.org
}
\description{
secopin
}
\details{
Question Based on what the doctor has just told you, would you seek a second opinion from another physician?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[3]} maybe
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2759/vshow}{GSS Data Explorer page for \code{secopin}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab maybe \tab don't know \tab no answer \tab Total \cr
   2002 \tab - \tab 1827 \tab 635 \tab 278 \tab 21 \tab 4 \tab 2765 \cr
   Total \tab 0 \tab 1827 \tab 635 \tab 278 \tab 21 \tab 4 \tab 2765 \cr
}
}

\seealso{
Other Doctors & Patients: 
\code{\link{doccosts}},
\code{\link{docinfo}},
\code{\link{docnomri}},
\code{\link{docsat}},
\code{\link{docswtch}},
\code{\link{insrchng}}
}
\concept{Doctors & Patients}
\keyword{variable}
