% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_569.R
\name{secprvcy}
\alias{secprvcy}
\title{Govt favor secrecy over workers privacy?}
\source{
General Social Survey https://gss.norc.org
}
\description{
secprvcy
}
\details{
Question When faced with the conflict between an individual''s right to privacy and the government''s need to collect personal
background information before giving a SECRET or TOP SECRET clearance, the government should:
\enumerate{
\item Favor protecting an individual''s right to privacy by not gathering information.
\item Favor the government''s need to protect security by gathering information.
}
}
\section{Values}{

\itemize{
\item \code{1} favor indiv right
\item \code{2} favor govt right
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5218/vshow}{GSS Data Explorer page for \code{secprvcy}}.
}

\seealso{
Other National Security: 
\code{\link{askcomp}},
\code{\link{askdrink}},
\code{\link{askfrbiz}},
\code{\link{askfrtrv}},
\code{\link{asksexor}},
\code{\link{secbudgt}},
\code{\link{secdiplo}},
\code{\link{secmilop}},
\code{\link{secterr}}

Other United States: 
\code{\link{askcomp}},
\code{\link{askdrink}},
\code{\link{askfrbiz}},
\code{\link{askfrtrv}},
\code{\link{asksexor}}
}
\concept{National Security}
\concept{United States}
\keyword{variable}
