% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_568.R
\name{sectech}
\alias{sectech}
\title{Should govt maintain secrecy over technology}
\source{
General Social Survey https://gss.norc.org
}
\description{
sectech
}
\details{
Question In order to maintain America's leadership in the world, the government should maintain a high level of secrecy surrounding technology with military uses. Do you . . .
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 1996, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5205/vshow}{GSS Data Explorer page for \code{sectech}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 501 \tab 601 \tab 169 \tab 105 \tab 45 \tab 47 \tab 6 \tab 2992 \cr
   1996 \tab 1444 \tab 455 \tab 579 \tab 142 \tab 171 \tab 60 \tab 46 \tab 7 \tab 2904 \cr
   1998 \tab 1445 \tab 446 \tab 511 \tab 160 \tab 152 \tab 51 \tab 59 \tab 8 \tab 2832 \cr
   Total \tab 4407 \tab 1402 \tab 1691 \tab 471 \tab 428 \tab 156 \tab 152 \tab 21 \tab 8728 \cr
}
}

\seealso{
Other National Security: 
\code{\link{askcrime}},
\code{\link{askdrugs}},
\code{\link{askfinan}},
\code{\link{askforgn}},
\code{\link{askmentl}},
\code{\link{rptcowrk}},
\code{\link{secdocs}}
}
\concept{National Security}
\keyword{variable}
