% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_569.R
\name{secterr}
\alias{secterr}
\title{Government maintain secrecy on domestic terrorism}
\source{
General Social Survey https://gss.norc.org
}
\description{
secterr
}
\details{
Question Should or should not the government maintain a high level of secrecy surrounding: (READ A-D)
C. Efforts to control domestic terrorism
}
\section{Values}{

\itemize{
\item \code{1} definitely should
\item \code{2} probably should
\item \code{3} probably should not
\item \code{4} definitely should not
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5221/vshow}{GSS Data Explorer page for \code{secterr}}.
}

\seealso{
Other National Security: 
\code{\link{askcomp}},
\code{\link{askdrink}},
\code{\link{askfrbiz}},
\code{\link{askfrtrv}},
\code{\link{asksexor}},
\code{\link{secbudgt}},
\code{\link{secdiplo}},
\code{\link{secmilop}},
\code{\link{secprvcy}}

Other Federal Government: 
\code{\link{secbudgt}},
\code{\link{secdiplo}},
\code{\link{secmilop}}
}
\concept{Federal Government}
\concept{National Security}
\concept{Terrorism}
\keyword{variable}
