% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_12.R
\name{sector}
\alias{sector}
\title{Type of college respondent attended}
\source{
General Social Survey https://gss.norc.org
}
\description{
sector
}
\details{
Question In what year did you receive the (COLLEGE NAME) degree?
A. Sector of Institution?
}
\section{Values}{

\itemize{
\item \code{0} administrative unit
\item \code{1} public, 4-year or above
\item \code{2} private not-for-profit, 4-year or above
\item \code{3} private for-profit, 4-year or above
\item \code{4} public, 2-year
\item \code{5} private not-for-profit, 2-year or less
\item \code{6} private for-profit, 2-year
\item \code{7} public, less than 2-year
\item \code{9} private for-profit, less than 2-year
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/69/vshow}{GSS Data Explorer page for \code{sector}}.
}

\seealso{
Other Core: 
\code{\link{barate}},
\code{\link{eftotlt}},
\code{\link{gradtounder}},
\code{\link{speftotlt}},
\code{\link{spsector}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme1}},
\code{\link{voednme2}}
}
\concept{Core}
\keyword{variable}
