% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_321.R
\name{seecowrk}
\alias{seecowrk}
\title{How easy to see whether co-workers working}
\source{
General Social Survey https://gss.norc.org
}
\description{
seecowrk
}
\details{
Question In your job how easy is it for you to see whether your co-workers are working well or poorly? On a scale of 0 to 10 please describe with 0 meaning not at all easy to see and 10 meaning very easy to see.
}
\section{Values}{

\itemize{
\item \verb{[0]} not at all easy to see
\item \verb{[10]} very easy to see
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2863/vshow}{GSS Data Explorer page for \code{seecowrk}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab not at all easy to see \tab 1 \tab very easy to see \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2002 \tab 969 \tab 194 \tab 30 \tab 755 \tab 31 \tab 35 \tab 44 \tab 106 \tab 53 \tab 116 \tab 239 \tab 138 \tab 36 \tab 19 \tab 2765 \cr
   2006 \tab 2778 \tab 145 \tab 20 \tab 793 \tab 31 \tab 38 \tab 38 \tab 114 \tab 39 \tab 91 \tab 212 \tab 116 \tab 65 \tab 30 \tab 4510 \cr
   Total \tab 3747 \tab 339 \tab 50 \tab 1548 \tab 62 \tab 73 \tab 82 \tab 220 \tab 92 \tab 207 \tab 451 \tab 254 \tab 101 \tab 49 \tab 7275 \cr
}
}

\seealso{
Other Employee Compensation: 
\code{\link{compperf}},
\code{\link{deptperf}},
\code{\link{extr2001}},
\code{\link{extrapay}},
\code{\link{extraval}},
\code{\link{extrayr}},
\code{\link{indperf}},
\code{\link{yearval}}
}
\concept{Employee Compensation}
\keyword{variable}
