% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_579.R
\name{seefilm}
\alias{seefilm}
\title{Last week went to see a film?}
\source{
General Social Survey https://gss.norc.org
}
\description{
seefilm
}
\details{
Question Now I'm going to ask you about things you did during the last seven days. I'm only interested in what you did during the last seven days.
From last (DAY OF WEEK) to today did you…
C. Go to a movie theater to see a film?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/C/- \tab partial \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5308/vshow}{GSS Data Explorer page for \code{seefilm}}.
}

\seealso{
Other Religious Attendance and Identity: 
\code{\link{attrelig}},
\code{\link{eatout}},
\code{\link{godoc}},
\code{\link{monday}},
\code{\link{numdays}},
\code{\link{sunday}}
}
\concept{Religious Attendance and Identity}
\keyword{variable}
