% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_367.R
\name{seeksci}
\alias{seeksci}
\title{Probable source of information about scientific issues}
\source{
General Social Survey https://gss.norc.org
}
\description{
seeksci
}
\details{
Question If you wanted to learn about scientific issues such as global warming or biotechnology, where would you get information (– newspapers, magazines, the Internet, books or other printed materials, TV, radio, government agencies, family, friends, colleagues, or some other source)?
}
\section{Values}{

\itemize{
\item \verb{[1]} newspapers
\item \verb{[2]} magazines
\item \verb{[3]} the internet
\item \verb{[4]} books/other printed material
\item \verb{[5]} tv
\item \verb{[6]} radio
\item \verb{[7]} government agencies
\item \verb{[8]} family
\item \verb{[9]} friends/colleagues
\item \verb{[10]} library
\item \verb{[11]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016, 2018 \cr
   A/B/C \tab 2008 \cr
   B/C/- \tab 2006, 2012, 2014 \cr
   C/-/- \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3316/vshow}{GSS Data Explorer page for \code{seeksci}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab newspapers \tab magazines \tab the internet \tab books/other printed material \tab tv \tab radio \tab government agencies \tab family \tab friends/colleagues \tab other \tab don't know \tab no answer \tab library \tab Total \cr
   2006 \tab 2646 \tab 117 \tab 108 \tab 921 \tab 181 \tab 398 \tab 21 \tab 4 \tab 23 \tab 16 \tab 38 \tab 35 \tab 2 \tab - \tab 4510 \cr
   2008 \tab 518 \tab 87 \tab 73 \tab 791 \tab 117 \tab 324 \tab 15 \tab 6 \tab 19 \tab 14 \tab 8 \tab 32 \tab 2 \tab 17 \tab 2023 \cr
   2010 \tab 1761 \tab 12 \tab 12 \tab 157 \tab 40 \tab 40 \tab 3 \tab 9 \tab 3 \tab 2 \tab - \tab 3 \tab - \tab 2 \tab 2044 \cr
   2012 \tab 972 \tab 34 \tab 40 \tab 599 \tab 68 \tab 190 \tab 13 \tab 10 \tab 11 \tab 5 \tab 1 \tab 25 \tab 4 \tab 2 \tab 1974 \cr
   2014 \tab 1299 \tab 51 \tab 39 \tab 789 \tab 76 \tab 191 \tab 23 \tab 25 \tab 10 \tab 17 \tab 6 \tab 8 \tab - \tab 4 \tab 2538 \cr
   2016 \tab 1477 \tab 42 \tab 45 \tab 927 \tab 101 \tab 175 \tab 16 \tab 24 \tab 19 \tab 11 \tab - \tab 15 \tab 5 \tab 10 \tab 2867 \cr
   2018 \tab 1173 \tab 28 \tab 25 \tab 791 \tab 95 \tab 145 \tab 14 \tab 20 \tab 22 \tab 11 \tab - \tab 14 \tab - \tab 10 \tab 2348 \cr
   Total \tab 9846 \tab 371 \tab 342 \tab 4975 \tab 678 \tab 1463 \tab 105 \tab 98 \tab 107 \tab 76 \tab 53 \tab 132 \tab 13 \tab 45 \tab 18304 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{advfront}},
\code{\link{bettrlfe}},
\code{\link{leadsci}},
\code{\link{newsfrom}},
\code{\link{nextgen}},
\code{\link{scifrom}},
\code{\link{scispec}},
\code{\link{toofast}},
\code{\link{whichsci}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
