% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_386.R
\name{seeksciy}
\alias{seeksciy}
\title{Where get science issue news}
\source{
General Social Survey https://gss.norc.org
}
\description{
seeksciy
}
\details{
Question If you wanted to learn about scientific issues such as global warming or biotechnology, where would you get information – printed newspapers, online newspapers, printed magazines, online magazines, other places on the Internet, books or other printed materials, TV, radio, family, friends, colleagues, or some other source)?
PROBE: What would be your main source of information?
}
\section{Values}{

\itemize{
\item \verb{[1]} print newspapers
\item \verb{[2]} online newspapers
\item \verb{[3]} print magazines
\item \verb{[4]} online magazines
\item \verb{[5]} other places on the internet
\item \verb{[6]} books/other printed material
\item \verb{[7]} tv
\item \verb{[8]} radio
\item \verb{[9]} family
\item \verb{[10]} friends/colleagues
\item \verb{[11]} other
\item \verb{[98]} don~~t know
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3505/vshow}{GSS Data Explorer page for \code{seeksciy}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab print newspapers \tab online newspapers \tab print magazines \tab online magazines \tab other places on the internet \tab books/other printed material \tab tv \tab radio \tab family \tab friends/colleagues \tab other \tab don't know \tab no answer \tab Total \cr
   2010 \tab 1386 \tab 19 \tab 45 \tab 15 \tab 32 \tab 297 \tab 111 \tab 93 \tab 8 \tab 9 \tab 7 \tab 15 \tab 4 \tab 3 \tab 2044 \cr
   Total \tab 1386 \tab 19 \tab 45 \tab 15 \tab 32 \tab 297 \tab 111 \tab 93 \tab 8 \tab 9 \tab 7 \tab 15 \tab 4 \tab 3 \tab 2044 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{cloning}},
\code{\link{newsfrmy}},
\code{\link{scientda}},
\code{\link{scientsn}},
\code{\link{scifromy}},
\code{\link{sciinfgo}},
\code{\link{scinews1}},
\code{\link{scinews2}},
\code{\link{scinews3}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
