% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_392.R
\name{seemhpub}
\alias{seemhpub}
\title{How frequently r sees someone mentally ill in public}
\source{
General Social Survey https://gss.norc.org
}
\description{
seemhpub
}
\details{
Question How frequently are you in a public place where you see someone who seems to be mentally ill…
}
\section{Values}{

\itemize{
\item \code{1} often
\item \code{2} sometimes
\item \code{3} almost never
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3568/vshow}{GSS Data Explorer page for \code{seemhpub}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab often \tab sometimes \tab almost never \tab never \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3073 \tab 329 \tab 468 \tab 460 \tab 148 \tab 20 \tab 12 \tab 4510 \cr
   Total \tab 3073 \tab 329 \tab 468 \tab 460 \tab 148 \tab 20 \tab 12 \tab 4510 \cr
}
}

\seealso{
Other Mental Health III: 
\code{\link{mhothrel}},
\code{\link{mhothyou}},
\code{\link{mhtrtoth}},
\code{\link{viglab06}}
}
\concept{Mental Health III}
\keyword{variable}
