% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_241.R
\name{seenmntl}
\alias{seenmntl}
\title{You or anyone you know seen a psychiatrist}
\source{
General Social Survey https://gss.norc.org
}
\description{
seenmntl
}
\details{
Question Have you or has anyone else you know ever seen a psychiatrist, psychologist, or counselor?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2167/vshow}{GSS Data Explorer page for \code{seenmntl}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 808 \tab 561 \tab 16 \tab 2 \tab 2832 \cr
   Total \tab 1445 \tab 808 \tab 561 \tab 16 \tab 2 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{psycmed1}},
\code{\link{psycmed2}},
\code{\link{psycmed3}}
}
\concept{Mental Health}
\keyword{variable}
