% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_241.R
\name{seenmntl}
\alias{seenmntl}
\title{You or anyone you know seen a psychiatrist}
\source{
General Social Survey https://gss.norc.org
}
\description{
seenmntl
}
\details{
Question Have you or has anyone else you know ever seen a psychiatrist, psychologist, or counselor?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2167/vshow}{GSS Data Explorer page for \code{seenmntl}}.
}

\seealso{
Other Mental Health: 
\code{\link{psycmed1}},
\code{\link{psycmed2}},
\code{\link{psycmed3}}

Other Mental Health: 
\code{\link{psycmed1}},
\code{\link{psycmed2}},
\code{\link{psycmed3}}
}
\concept{Mental Health}
\keyword{variable}
