% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_384.R
\name{seesand}
\alias{seesand}
\title{Which picture shows what she should use for the 2nd pot}
\source{
General Social Survey https://gss.norc.org
}
\description{
seesand
}
\details{
Question Please look at Card I-4. A gardener has an idea that a plant needs sand in the soil for healthy growth. In order to test her idea she uses two pots of plants. She sets up one pot of plants as shown on the top part of the card. Which one of the pictures on the bottom part of the card shows what should she use for the second pot?
}
\section{Values}{

\itemize{
\item \verb{[1]} sunlight. sand and water
\item \verb{[2]} dark cupboard. sand, soil, and water
\item \verb{[3]} dark cupboard. soil and water
\item \verb{[4]} sunlight. sand and soil
\item \verb{[5]} sunlight. soil and water
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3484/vshow}{GSS Data Explorer page for \code{seesand}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab sunlight. sand and water \tab dark cupboard. sand, soil, and water \tab dark cupboard. soil and water \tab sunlight. sand and soil \tab sunlight. soil and water \tab don't know \tab no answer \tab Total \cr
   2008 \tab 709 \tab 105 \tab 170 \tab 43 \tab 144 \tab 692 \tab 136 \tab 24 \tab 2023 \cr
   Total \tab 709 \tab 105 \tab 170 \tab 43 \tab 144 \tab 692 \tab 136 \tab 24 \tab 2023 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{daynight}},
\code{\link{fishexp1}},
\code{\link{fishexp2}},
\code{\link{genes}},
\code{\link{gills}},
\code{\link{morempg}},
\code{\link{polnuke}},
\code{\link{upbreath}},
\code{\link{weighing}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
