% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_223.R
\name{seetalk2}
\alias{seetalk2}
\title{How frequently r sees or talks to person 2}
\source{
General Social Survey https://gss.norc.org
}
\description{
seetalk2
}
\details{
Question On a scale from 1 to 10, with 1 being not at all and 10 being very often, how frequently do you see or talk to (NAME 2) these days?
}
\section{Values}{

\itemize{
\item \code{1} 1 - not at all
\item \code{10} 10 - very often
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7116/vshow}{GSS Data Explorer page for \code{seetalk2}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 1 - not at all \tab 10 - very often \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2018 \tab 1982 \tab 99 \tab 59 \tab 31 \tab 34 \tab 24 \tab 29 \tab 28 \tab 13 \tab 29 \tab 12 \tab 4 \tab 4 \tab 2348 \cr
   Total \tab 1982 \tab 99 \tab 59 \tab 31 \tab 34 \tab 24 \tab 29 \tab 28 \tab 13 \tab 29 \tab 12 \tab 4 \tab 4 \tab 2348 \cr
}
}

\keyword{variable}
