% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_506.R
\name{selfemp1}
\alias{selfemp1}
\title{Employee have more job security than the self-employed}
\source{
General Social Survey https://gss.norc.org
}
\description{
selfemp1
}
\details{
Question To what extent do you agree or disagree with the following statements?
A. Employees have more job security than the self-employed
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4606/vshow}{GSS Data Explorer page for \code{selfemp1}}.
}

\seealso{
Other Work Orientation: 
\code{\link{adforjob}},
\code{\link{appemps}},
\code{\link{askhelp}},
\code{\link{econsup}},
\code{\link{rimpskls}},
\code{\link{selfemp2}},
\code{\link{unbetter}},
\code{\link{unjobsec}},
\code{\link{wantads}}

Other Work: 
\code{\link{rimpskls}},
\code{\link{selfemp2}},
\code{\link{unbetter}},
\code{\link{unjobsec}}
}
\concept{Work}
\concept{Work Orientation}
\keyword{variable}
