% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_324.R
\name{selffrst}
\alias{selffrst}
\title{People need not overly worry about others}
\source{
General Social Survey https://gss.norc.org
}
\description{
selffrst
}
\details{
Question Please tell me whether you strongly agree, agree, neither agree nor disagree, disagree, or strongly disagree with the following statements:
D. These days people need to look after themselves and not overly worry about others
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004, 2014 \cr
   A/C/- \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2893/vshow}{GSS Data Explorer page for \code{selffrst}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1393 \tab 86 \tab 358 \tab 289 \tab 508 \tab 114 \tab 11 \tab 6 \tab 2765 \cr
   2004 \tab 1472 \tab 59 \tab 265 \tab 322 \tab 573 \tab 115 \tab 4 \tab 2 \tab 2812 \cr
   2012 \tab 666 \tab 61 \tab 358 \tab 274 \tab 521 \tab 77 \tab 8 \tab 9 \tab 1974 \cr
   2014 \tab 1264 \tab 72 \tab 326 \tab 303 \tab 478 \tab 82 \tab 6 \tab 7 \tab 2538 \cr
   Total \tab 4795 \tab 278 \tab 1307 \tab 1188 \tab 2080 \tab 388 \tab 29 \tab 24 \tab 10089 \cr
}
}

\seealso{
Other Altruism: 
\code{\link{accptoth}},
\code{\link{careself}},
\code{\link{carried}},
\code{\link{directns}},
\code{\link{helpaway}},
\code{\link{loanitem}},
\code{\link{othshelp}},
\code{\link{peoptrbl}},
\code{\link{selfless}}
}
\concept{Altruism}
\concept{Gender}
\keyword{variable}
