% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_146.R
\name{selfirst}
\alias{selfirst}
\title{Take care of yourself first before others}
\source{
General Social Survey https://gss.norc.org
}
\description{
selfirst
}
\details{
Question As I read each of the following statements, please tell me whether you strongly agree, agree, neither agree or disagree, disagree, or strongly disagree with it.
A. You have to take care of yourself first, and if you have any energy left over, then help other people.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1377/vshow}{GSS Data Explorer page for \code{selfirst}}.
}

\seealso{
Other Culture: 
\code{\link{richpoor}},
\code{\link{standup}}
}
\concept{Anomia}
\concept{Culture}
\keyword{variable}
