% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_190.R
\name{selfresp}
\alias{selfresp}
\title{How resp. do you feel for the situation?}
\source{
General Social Survey https://gss.norc.org
}
\description{
selfresp
}
\details{
Question How responsible do YOU feel for the situation? Think of a scale where 0 means you were not at all responsible and 10 means you were totally responsible for what happened.
}
\section{Values}{

\itemize{
\item \verb{[0]} not at all responsible
\item \verb{[10]} completely responsible
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1785/vshow}{GSS Data Explorer page for \code{selfresp}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab not at all responsible \tab 1 \tab completely responsible \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab no answer \tab Total \cr
   1996 \tab 1908 \tab 443 \tab 92 \tab 31 \tab 90 \tab 78 \tab 50 \tab 96 \tab 25 \tab 22 \tab 27 \tab 10 \tab 32 \tab 2904 \cr
   Total \tab 1908 \tab 443 \tab 92 \tab 31 \tab 90 \tab 78 \tab 50 \tab 96 \tab 25 \tab 22 \tab 27 \tab 10 \tab 32 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{amimp}},
\code{\link{chnang16}},
\code{\link{likeoth}},
\code{\link{notam}},
\code{\link{othresp}},
\code{\link{reactok}},
\code{\link{usfrustr}},
\code{\link{ussat}},
\code{\link{usworry}}
}
\concept{Emotions}
\keyword{variable}
