% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_190.R
\name{selfresp}
\alias{selfresp}
\title{How resp. do you feel for the situation?}
\source{
General Social Survey https://gss.norc.org
}
\description{
selfresp
}
\details{
Question How responsible do YOU feel for the situation? Think of a scale where 0 means you were not at all responsible and 10 means you were totally responsible for what happened.
}
\section{Values}{

\itemize{
\item \code{0} not at all responsible
\item \code{10} completely responsible
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1785/vshow}{GSS Data Explorer page for \code{selfresp}}.
}

\seealso{
Other Emotions: 
\code{\link{amimp}},
\code{\link{chnang16}},
\code{\link{likeoth}},
\code{\link{notam}},
\code{\link{othresp}},
\code{\link{reactok}},
\code{\link{usfrustr}},
\code{\link{ussat}},
\code{\link{usworry}}
}
\concept{Emotions}
\keyword{variable}
