% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_196.R
\name{sellbed}
\alias{sellbed}
\title{Would you rather sell bedroom furniture}
\source{
General Social Survey https://gss.norc.org
}
\description{
sellbed
}
\details{
Question I'd like you to imagine that you want to sell each of the following items. Would you rather sell these items to someone with whom you have had some family or social relationship, or to a buyer with whom you had no prior personal contact? Please tell me the number that best represents your feeling with respect to each of the sales listed.
A. Bedroom furniture.
}
\section{Values}{

\itemize{
\item \code{1} strong preference no contact
\item \code{3} dont care
\item \code{5} strong preference contact
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1851/vshow}{GSS Data Explorer page for \code{sellbed}}.
}

\seealso{
Other Markets: 
\code{\link{adoption}},
\code{\link{borrowed}},
\code{\link{buyauto}},
\code{\link{buybed}},
\code{\link{buyfixes}},
\code{\link{buyhome}},
\code{\link{buylaw}},
\code{\link{carbuy1}},
\code{\link{carbuya}},
\code{\link{carbuyb}},
\code{\link{carbuyc}},
\code{\link{carbuyd}},
\code{\link{carbuye}},
\code{\link{carbuyf}},
\code{\link{carbuyg}},
\code{\link{cardealr}},
\code{\link{carprivt}},
\code{\link{decauto}},
\code{\link{decbed}},
\code{\link{decgift}},
\code{\link{econsys}},
\code{\link{evbuyhme}},
\code{\link{fambiz}},
\code{\link{famfinan}},
\code{\link{fixer1}},
\code{\link{fixera}},
\code{\link{fixerb}},
\code{\link{fixerc}},
\code{\link{fixerd}},
\code{\link{fixere}},
\code{\link{fixerf}},
\code{\link{fixhome}},
\code{\link{frndawk}},
\code{\link{frnddeal}},
\code{\link{haggle}},
\code{\link{homeyear}},
\code{\link{hrfambiz}},
\code{\link{incgap}},
\code{\link{inequal1}},
\code{\link{inequal3}},
\code{\link{inequal5}},
\code{\link{lawyer}},
\code{\link{lawyer1}},
\code{\link{lawyera}},
\code{\link{lawyerb}},
\code{\link{lawyerc}},
\code{\link{lawyerd}},
\code{\link{lawyere}},
\code{\link{lawyerf}},
\code{\link{lawyerg}},
\code{\link{layoffs}},
\code{\link{learnhme}},
\code{\link{loaned}},
\code{\link{loanedto}},
\code{\link{natrecon}},
\code{\link{newowned}},
\code{\link{newused}},
\code{\link{organsb}},
\code{\link{organsw}},
\code{\link{realtor1}},
\code{\link{realtora}},
\code{\link{realtorb}},
\code{\link{realtorc}},
\code{\link{realtord}},
\code{\link{realtore}},
\code{\link{realtorf}},
\code{\link{relagent}},
\code{\link{reldealr}},
\code{\link{relfixer}},
\code{\link{relhome}},
\code{\link{rellaw}},
\code{\link{relprivt}},
\code{\link{reqinfo}},
\code{\link{satcar}},
\code{\link{satfixes}},
\code{\link{sathome}},
\code{\link{satlawyr}},
\code{\link{sellauto}},
\code{\link{sellbaby}},
\code{\link{sellhome}},
\code{\link{sellorgn}},
\code{\link{sellsex}},
\code{\link{tablprce}},
\code{\link{tagsales}},
\code{\link{transoth}},
\code{\link{transrel}},
\code{\link{typdealr}},
\code{\link{warrntyc}},
\code{\link{warrntyh}},
\code{\link{wholoand}},
\code{\link{whosold}},
\code{\link{wkfambiz}}
}
\concept{Markets}
\keyword{variable}
