% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_198.R
\name{sellorgn}
\alias{sellorgn}
\title{Should people be permitted to sell a kidney}
\source{
General Social Survey https://gss.norc.org
}
\description{
sellorgn
}
\details{
Question A body organ that is much in need and that people may contribute are kidneys. Most people can live with only one kidney, though their chances of survival are better if they have two. Do you believe that people with two healthy kidneys should be permitted to sell a kidney to a hospital or organ center to use for transplants?
}
\section{Values}{

\itemize{
\item \code{1} definitely not
\item \code{2} probably not
\item \code{3} perhaps
\item \code{4} probably
\item \code{5} definitely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1868/vshow}{GSS Data Explorer page for \code{sellorgn}}.
}

\seealso{
Other Markets: 
\code{\link{adoption}},
\code{\link{frndawk}},
\code{\link{frnddeal}},
\code{\link{organsb}},
\code{\link{organsw}},
\code{\link{sellbaby}},
\code{\link{sellsex}},
\code{\link{tablprce}},
\code{\link{transrel}}
}
\concept{Markets}
\keyword{variable}
