% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_198.R
\name{sellsex}
\alias{sellsex}
\title{Nothing inherently wrong with prostitution}
\source{
General Social Survey https://gss.norc.org
}
\description{
sellsex
}
\details{
Question How much do you agree or disagree with the following statements:
A. There is nothing inherently wrong with prostitution, so long as the health risks can be minimized. If consenting adults agree to exchange money for sex, that is their business.
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree somewhat
\item \code{3} disagree somewhat
\item \code{4} disagree strongly
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1871/vshow}{GSS Data Explorer page for \code{sellsex}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab agree strongly \tab agree somewhat \tab disagree somewhat \tab disagree strongly \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 366 \tab 291 \tab 270 \tab 478 \tab 35 \tab 4 \tab 2904 \cr
   Total \tab 1460 \tab 366 \tab 291 \tab 270 \tab 478 \tab 35 \tab 4 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{adoption}},
\code{\link{frndawk}},
\code{\link{frnddeal}},
\code{\link{organsb}},
\code{\link{organsw}},
\code{\link{sellbaby}},
\code{\link{sellorgn}},
\code{\link{tablprce}},
\code{\link{transrel}}
}
\concept{Markets}
\keyword{variable}
