% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_406.R
\name{semiguns}
\alias{semiguns}
\title{Should semi-auto gun sale be limited to military}
\source{
General Social Survey https://gss.norc.org
}
\description{
semiguns
}
\details{
Question Should semi-automatic, assault weapons or semi-automatic guns known as assault rifles be sold to the general public or should their sales be limited to the military and police?
}
\section{Values}{

\itemize{
\item \verb{[1]} sold to the general public
\item \verb{[2]} sales be limited to the military and police
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3702/vshow}{GSS Data Explorer page for \code{semiguns}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab sold to the general public \tab sales be limited to the military and police \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3139 \tab 198 \tab 1125 \tab 36 \tab 12 \tab 4510 \cr
   Total \tab 3139 \tab 198 \tab 1125 \tab 36 \tab 12 \tab 4510 \cr
}
}

\seealso{
Other Gun Control: 
\code{\link{guns911}},
\code{\link{gunsales}},
\code{\link{gunsdrnk}},
\code{\link{gunsdrug}},
\code{\link{othguns}},
\code{\link{rifles50}}
}
\concept{Gun Control}
\keyword{variable}
