% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_210.R
\name{seriousp}
\alias{seriousp}
\title{How serious x s situation is}
\source{
General Social Survey https://gss.norc.org
}
\description{
seriousp
}
\details{
Question Please remember, there are no right or wrong answers to these questions. Please think about the person I just described when answering this group of questions. First, how serious would you consider NAME's problem to be -- very serious, somewhat serious, not very serious, or not at all serious?
}
\section{Values}{

\itemize{
\item \code{1} very serious
\item \code{2} somewhat serious
\item \code{3} not very serious
\item \code{4} not at all serious
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1987/vshow}{GSS Data Explorer page for \code{seriousp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very serious \tab somewhat serious \tab not very serious \tab not at all serious \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 899 \tab 308 \tab 148 \tab 65 \tab 9 \tab 15 \tab 2904 \cr
   2006 \tab 3073 \tab 828 \tab 349 \tab 180 \tab 53 \tab 13 \tab 14 \tab 4510 \cr
   2018 \tab 1175 \tab 720 \tab 265 \tab 122 \tab 34 \tab 4 \tab 28 \tab 2348 \cr
   Total \tab 5708 \tab 2447 \tab 922 \tab 450 \tab 152 \tab 26 \tab 57 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{charactr}},
\code{\link{imbalnce}},
\code{\link{probfix3}},
\code{\link{probhlp1}},
\code{\link{probhlp2}},
\code{\link{probhlp3}},
\code{\link{stresses}},
\code{\link{vigversn}},
\code{\link{wayraise}}
}
\concept{Mental Health}
\keyword{variable}
