% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_210.R
\name{seriousp}
\alias{seriousp}
\title{How serious x s situation is}
\source{
General Social Survey https://gss.norc.org
}
\description{
seriousp
}
\details{
Question Please remember, there are no right or wrong answers to these questions. Please think about the person I just described when answering this group of questions. First, how serious would you consider NAME's problem to be -- very serious, somewhat serious, not very serious, or not at all serious?
}
\section{Values}{

\itemize{
\item \code{1} very serious
\item \code{2} somewhat serious
\item \code{3} not very serious
\item \code{4} not at all serious
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2018 \tab B/C/- \tab partial \cr
   2024 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1987/vshow}{GSS Data Explorer page for \code{seriousp}}.
}

\seealso{
Other Mental Health: 
\code{\link{charactr}},
\code{\link{imbalnce}},
\code{\link{probfix3}},
\code{\link{probhlp1}},
\code{\link{probhlp2}},
\code{\link{probhlp3}},
\code{\link{stresses}},
\code{\link{vigversn}},
\code{\link{wayraise}}

Other Mental Health: 
\code{\link{charactr}},
\code{\link{imbalnce}},
\code{\link{probfix3}},
\code{\link{probhlp1}},
\code{\link{probhlp2}},
\code{\link{probhlp3}},
\code{\link{stresses}},
\code{\link{vigversn}},
\code{\link{wayraise}}
}
\concept{Mental Health}
\keyword{variable}
