% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_542.R
\name{servepeo}
\alias{servepeo}
\title{How committed govt admnstrator's are to serve people}
\source{
General Social Survey https://gss.norc.org
}
\description{
servepeo
}
\details{
Question Thinking of government administrators in America, how committed are they to serve the people?
}
\section{Values}{

\itemize{
\item \code{1} very committed
\item \code{2} somewhat committed
\item \code{3} not very committed
\item \code{4} not at all committed
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4957/vshow}{GSS Data Explorer page for \code{servepeo}}.
}

\seealso{
Other ISSP Citizenship: 
\code{\link{chngeoth}},
\code{\link{choices}},
\code{\link{elecfair}},
\code{\link{elecvote}},
\code{\link{govngos}},
\code{\link{polactve}},
\code{\link{powerun}},
\code{\link{refrndms}},
\code{\link{unrghts}}
}
\concept{ISSP Citizenship}
\keyword{variable}
