% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_542.R
\name{servepeo}
\alias{servepeo}
\title{How committed govt admnstrator's are to serve people}
\source{
General Social Survey https://gss.norc.org
}
\description{
servepeo
}
\details{
Question Thinking of government administrators in America, how committed are they to serve the people?
}
\section{Values}{

\itemize{
\item \verb{[1]} very committed
\item \verb{[2]} somewhat committed
\item \verb{[3]} not very committed
\item \verb{[4]} not at all committed
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4957/vshow}{GSS Data Explorer page for \code{servepeo}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very committed \tab somewhat committed \tab not very committed \tab not at all committed \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 167 \tab 923 \tab 286 \tab 69 \tab 21 \tab 6 \tab 2812 \cr
   2014 \tab 1274 \tab 94 \tab 658 \tab 334 \tab 102 \tab 70 \tab 6 \tab 2538 \cr
   Total \tab 2614 \tab 261 \tab 1581 \tab 620 \tab 171 \tab 91 \tab 12 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{chngeoth}},
\code{\link{choices}},
\code{\link{elecfair}},
\code{\link{elecvote}},
\code{\link{govngos}},
\code{\link{polactve}},
\code{\link{powerun}},
\code{\link{refrndms}},
\code{\link{unrghts}}
}
\concept{ISSP Citizenship}
\keyword{variable}
