% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_680.R
\name{servpub}
\alias{servpub}
\title{How much the public service serves people}
\source{
General Social Survey https://gss.norc.org
}
\description{
servpub
}
\details{
Question Thinking of the public service in America, how committed is it to serve the people?
}
\section{Values}{

\itemize{
\item \code{1} very committed
\item \code{2} somewhat committed
\item \code{3} not very committed
\item \code{4} not at all committed
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8168/vshow}{GSS Data Explorer page for \code{servpub}}.
}

\seealso{
Other ISSP Citizenship: 
\code{\link{leftrght1}},
\code{\link{poltalk}},
\code{\link{usmca1}}
}
\concept{ISSP Citizenship}
\keyword{variable}
