% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_679.R
\name{servpub}
\alias{servpub}
\title{How much the public service serves people}
\source{
General Social Survey https://gss.norc.org
}
\description{
servpub
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} very committed
\item \code{2} somewhat committed
\item \code{3} not very committed
\item \code{4} not at all committed
\item \code{NA(d)} can't choose
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
