% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_441.R
\name{setwage}
\alias{setwage}
\title{Govt control of wages}
\source{
General Social Survey https://gss.norc.org
}
\description{
setwage
}
\details{
Question Here are some things the government might do for the economy. Circle one number for each action to show whether you are in favor of it or against it.
A. Control of wages by legislation.
}
\section{Values}{

\itemize{
\item \code{1} strongly in favor
\item \code{2} in favor
\item \code{3} neither
\item \code{4} against
\item \code{5} strongly against
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4049/vshow}{GSS Data Explorer page for \code{setwage}}.
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{cutgovt}},
\code{\link{cuthours}},
\code{\link{hlphitec}},
\code{\link{lessreg}},
\code{\link{makejobs}},
\code{\link{poleff10}},
\code{\link{savejobs}},
\code{\link{setprice}},
\code{\link{spenviro}}

Other Economy: 
\code{\link{cutgovt}},
\code{\link{cuthours}},
\code{\link{hlphitec}},
\code{\link{lessreg}},
\code{\link{makejobs}},
\code{\link{savejobs}},
\code{\link{setprice}}

Other Wage/Price Controls: 
\code{\link{setprice}}
}
\concept{Economy}
\concept{ISSP Role of Gov't}
\concept{Wage/Price Controls}
\keyword{variable}
