% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_95.R
\name{sex2}
\alias{sex2}
\title{Sex of person number 2}
\source{
General Social Survey https://gss.norc.org
}
\description{
sex2
}
\details{
Question (NAME) is (male/female)?  Is that correct?  ASK FOR EACH NAME.
2. Sex of Second Person
}
\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/866/vshow}{GSS Data Explorer page for \code{sex2}}.
}

\seealso{
Other Social Networks: 
\code{\link{close24}},
\code{\link{close25}},
\code{\link{close34}},
\code{\link{close35}},
\code{\link{close45}},
\code{\link{sex1}},
\code{\link{sex3}},
\code{\link{sex4}},
\code{\link{sex5}}

Other Sex: 
\code{\link{sex1}},
\code{\link{sex3}},
\code{\link{sex4}},
\code{\link{sex5}}

Other Social Networks: 
\code{\link{close24}},
\code{\link{close25}},
\code{\link{close34}},
\code{\link{close35}},
\code{\link{close45}},
\code{\link{sex1}},
\code{\link{sex3}},
\code{\link{sex4}},
\code{\link{sex5}}
}
\concept{Sex}
\concept{Social Networks}
\keyword{variable}
