% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_95.R
\name{sex3}
\alias{sex3}
\title{Sex of person number 3}
\source{
General Social Survey https://gss.norc.org
}
\description{
sex3
}
\details{
Question (NAME) is (male/female)?  Is that correct?  ASK FOR EACH NAME.
3. Sex of Third Person
}
\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/867/vshow}{GSS Data Explorer page for \code{sex3}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab male \tab female \tab no answer \tab Total \cr
   1985 \tab 599 \tab 447 \tab 485 \tab 3 \tab 1534 \cr
   2004 \tab 2240 \tab 228 \tab 297 \tab 47 \tab 2812 \cr
   Total \tab 2839 \tab 675 \tab 782 \tab 50 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{close24}},
\code{\link{close25}},
\code{\link{close34}},
\code{\link{close35}},
\code{\link{close45}},
\code{\link{sex1}},
\code{\link{sex2}},
\code{\link{sex4}},
\code{\link{sex5}}
}
\concept{Social Networks}
\keyword{variable}
