% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_95.R
\name{sex4}
\alias{sex4}
\title{Sex of person number 4}
\source{
General Social Survey https://gss.norc.org
}
\description{
sex4
}
\details{
Question (NAME) is (male/female)?  Is that correct?  ASK FOR EACH NAME.
4. Sex of Fourth Person
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/868/vshow}{GSS Data Explorer page for \code{sex4}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab male \tab female \tab no answer \tab Total \cr
   1985 \tab 920 \tab 292 \tab 319 \tab 3 \tab 1534 \cr
   2004 \tab 2472 \tab 137 \tab 157 \tab 46 \tab 2812 \cr
   Total \tab 3392 \tab 429 \tab 476 \tab 49 \tab 4346 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{close24}},
\code{\link{close25}},
\code{\link{close34}},
\code{\link{close35}},
\code{\link{close45}},
\code{\link{sex1}},
\code{\link{sex2}},
\code{\link{sex3}},
\code{\link{sex5}}
}
\concept{Social Networks}
\keyword{variable}
