% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_555.R
\name{sexbirth}
\alias{sexbirth}
\title{What sex were you assigned at birth?  (For example, on your birth certificate.)}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexbirth
}
\details{
Question What sex were you assigned at birth? (For example, on your birth certificate)
}
\section{Values}{

\itemize{
\item \code{1} female
\item \code{2} male
\item \code{3} intersex
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7292/vshow}{GSS Data Explorer page for \code{sexbirth}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab female \tab male \tab intersex \tab no answer \tab Total \cr
   2018 \tab 939 \tab 759 \tab 637 \tab 1 \tab 12 \tab 2348 \cr
   Total \tab 939 \tab 759 \tab 637 \tab 1 \tab 12 \tab 2348 \cr
}
}

\keyword{variable}
