% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_627.R
\name{sexbirth1}
\alias{sexbirth1}
\title{R's sex assigned at birth (2021)}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexbirth1
}
\details{
Question Was your sex recorded as male or female at birth?
}
\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7478/vshow}{GSS Data Explorer page for \code{sexbirth1}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab male \tab female \tab don't know \tab skipped on web \tab no answer \tab Total \cr
   2021 \tab 71 \tab 1730 \tab 2198 \tab 21 \tab 12 \tab - \tab 4032 \cr
   2022 \tab 8 \tab 1625 \tab 1894 \tab 12 \tab 4 \tab 1 \tab 3544 \cr
   Total \tab 79 \tab 3355 \tab 4092 \tab 33 \tab 16 \tab 1 \tab 7576 \cr
}
}

\keyword{variable}
