% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_553.R
\name{sexfreq1}
\alias{sexfreq1}
\title{Had sex in last month}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexfreq1
}
\details{
Question In the past month, have you engaged in sexual intercourse?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5058/vshow}{GSS Data Explorer page for \code{sexfreq1}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no \tab Total \cr
   1990 \tab 791 \tab 388 \tab 193 \tab 1372 \cr
   Total \tab 791 \tab 388 \tab 193 \tab 1372 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{nummen}},
\code{\link{numwomen}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{sexfreq2}},
\code{\link{sexsex5}}
}
\concept{High Risk Behaviors}
\keyword{variable}
