% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_553.R
\name{sexfreq1}
\alias{sexfreq1}
\title{Had sex in last month}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexfreq1
}
\details{
Question In the past month, have you engaged in sexual intercourse?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5058/vshow}{GSS Data Explorer page for \code{sexfreq1}}.
}

\seealso{
Other High Risk Behaviors: 
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{nummen}},
\code{\link{numwomen}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{sexfreq2}},
\code{\link{sexsex5}}
}
\concept{High Risk Behaviors}
\keyword{variable}
