% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_553.R
\name{sexfreq2}
\alias{sexfreq2}
\title{Number of times, sex last month}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexfreq2
}
\details{
Question About how many times did you engage in intercourse during the month?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5059/vshow}{GSS Data Explorer page for \code{sexfreq2}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 19 \tab 2 \tab 20 \tab 22 \tab 23 \tab 24 \tab 25 \tab 27 \tab 3 \tab 30 \tab 4 \tab 5 \tab 50 \tab 6 \tab 60 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   1990 \tab 984 \tab 28 \tab 37 \tab 16 \tab 6 \tab 1 \tab 13 \tab 4 \tab 1 \tab 1 \tab 36 \tab 13 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 30 \tab 4 \tab 45 \tab 30 \tab 1 \tab 32 \tab 1 \tab 10 \tab 36 \tab 4 \tab 1 \tab 30 \tab 1372 \cr
   Total \tab 984 \tab 28 \tab 37 \tab 16 \tab 6 \tab 1 \tab 13 \tab 4 \tab 1 \tab 1 \tab 36 \tab 13 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 30 \tab 4 \tab 45 \tab 30 \tab 1 \tab 32 \tab 1 \tab 10 \tab 36 \tab 4 \tab 1 \tab 30 \tab 1372 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{evpaidsx}},
\code{\link{evstray}},
\code{\link{nummen}},
\code{\link{numwomen}},
\code{\link{partnrs5}},
\code{\link{partopen}},
\code{\link{partopn5}},
\code{\link{sexfreq1}},
\code{\link{sexsex5}}
}
\concept{High Risk Behaviors}
\keyword{variable}
