% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_484.R
\name{sexhar}
\alias{sexhar}
\title{R experienced sexual harassment}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexhar
}
\details{
Question Sometimes at work people find themselves the object of sexual advances, propositions, or unwanted sexual discussions from co-workers or supervisors. The advances sometimes involve physical contact and sometimes just involve sexual conversations. Has this ever happened to you?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[3]} never have worked
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4416/vshow}{GSS Data Explorer page for \code{sexhar}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab never have worked \tab no answer \tab don't know \tab Total \cr
   1994 \tab 1545 \tab 523 \tab 801 \tab 31 \tab 92 \tab - \tab 2992 \cr
   1996 \tab 1444 \tab 456 \tab 949 \tab - \tab 51 \tab 4 \tab 2904 \cr
   Total \tab 2989 \tab 979 \tab 1750 \tab 31 \tab 143 \tab 4 \tab 5896 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{abchoose}},
\code{\link{caresick}},
\code{\link{chldcare}},
\code{\link{fambudgt}},
\code{\link{laundry}},
\code{\link{livnowed}},
\code{\link{mapaid}},
\code{\link{repairs}},
\code{\link{teensex1}}
}
\concept{Gender}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
