% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_555.R
\name{sexnow}
\alias{sexnow}
\title{What is your current gender?}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexnow
}
\details{
Question What is your current gender?
}
\section{Values}{

\itemize{
\item \code{1} female
\item \code{2} male
\item \code{3} transgender
\item \code{4} a gender not listed here
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2018 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7293/vshow}{GSS Data Explorer page for \code{sexnow}}.
}

\keyword{variable}
