% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_627.R
\name{sexnow1}
\alias{sexnow1}
\title{R's sex now (2021)}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexnow1
}
\details{
Question Do you describe yourself as male, female, or transgender?
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[3]} transgender
\item \verb{[4]} none of these
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7479/vshow}{GSS Data Explorer page for \code{sexnow1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab male \tab female \tab transgender \tab none of these \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 72 \tab 1727 \tab 2179 \tab 10 \tab 16 \tab 18 \tab 2 \tab 8 \tab 4032 \cr
   2022 \tab 8 \tab 1617 \tab 1881 \tab 14 \tab 11 \tab - \tab 9 \tab 4 \tab 3544 \cr
   Total \tab 80 \tab 3344 \tab 4060 \tab 24 \tab 27 \tab 18 \tab 11 \tab 12 \tab 7576 \cr
}
}

\keyword{variable}
