% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_627.R
\name{sexnow1}
\alias{sexnow1}
\title{R's sex now (2021)}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexnow1
}
\details{
Question Do you describe yourself as male, female, or transgender?
}
\section{Values}{

\itemize{
\item \code{1} male
\item \code{2} female
\item \code{3} transgender
\item \code{4} none of these
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7479/vshow}{GSS Data Explorer page for \code{sexnow1}}.
}

\keyword{variable}
