% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_555.R
\name{sexornt}
\alias{sexornt}
\title{Sexual orientation}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexornt
}
\details{
Question Which of the following best describes you?
}
\section{Values}{

\itemize{
\item \code{1} gay, lesbian, or homosexual
\item \code{2} bisexual
\item \code{3} heterosexual or straight
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab B/C/- \tab partial \cr
   2018 \tab B/C/- \tab partial \cr
   2021 \tab B/C/- \tab partial \cr
   2022 \tab B/C/- \tab partial \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5081/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab bisexual \tab don't know \tab gay, lesbian, or homosexual \tab heterosexual or straight \tab no answer \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 236 \tab 27 \tab 7 \tab 33 \tab 1699 \tab 21 \tab - \tab 2023 \cr
   2010 \tab 202 \tab 31 \tab 9 \tab 27 \tab 1749 \tab 26 \tab - \tab 2044 \cr
   2012 \tab 237 \tab 43 \tab 3 \tab 27 \tab 1637 \tab 27 \tab - \tab 1974 \cr
   2014 \tab 188 \tab 65 \tab 6 \tab 45 \tab 2195 \tab 39 \tab - \tab 2538 \cr
   2016 \tab 1095 \tab 56 \tab 5 \tab 46 \tab 1641 \tab 24 \tab - \tab 2867 \cr
   2018 \tab 942 \tab 51 \tab 6 \tab 31 \tab 1291 \tab 27 \tab - \tab 2348 \cr
   2021 \tab 1682 \tab 96 \tab 53 \tab 76 \tab 2088 \tab 2 \tab 35 \tab 4032 \cr
   2022 \tab 1900 \tab 82 \tab 62 \tab 53 \tab 1409 \tab 7 \tab 31 \tab 3544 \cr
   2024 \tab - \tab 122 \tab 41 \tab 67 \tab 3052 \tab 11 \tab 16 \tab 3309 \cr
   Total \tab 57502 \tab 573 \tab 192 \tab 405 \tab 16761 \tab 184 \tab 82 \tab 75699 \cr
}
}

\seealso{
Other High Risk Behaviors: 
\code{\link{aidscndm}},
\code{\link{hivkiss}},
\code{\link{hivvac}}
}
\concept{Gays And Lesbians}
\concept{High Risk Behaviors}
\concept{Sex}
\keyword{variable}
