% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_682.R
\name{sexornt1}
\alias{sexornt1}
\title{Sexual orientation (reask, self-administered)}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexornt1
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} gay, lesbian, or homosexual
\item \code{2} bisexual
\item \code{3} heterosexual or straight
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
