% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_682.R
\name{sexornt1}
\alias{sexornt1}
\title{Sexual orientation (reask, self-administered)}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexornt1
}
\details{
Question Which of the following best describes you?
}
\section{Values}{

\itemize{
\item \code{1} gay, lesbian, or homosexual
\item \code{2} bisexual
\item \code{3} heterosexual or straight
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8171/vshow}{GSS Data Explorer page for \code{sexornt1}}.
}

\seealso{
Other High Risk Behaviors: 
\code{\link{sexpaid}}
}
\concept{High Risk Behaviors}
\keyword{variable}
