% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_682.R
\name{sexpaid}
\alias{sexpaid}
\title{Ever had sex with a person you paid or who paid you for sex}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexpaid
}
\details{
Question Have you ever had sex with a person  you paid or who paid you for sex?
As of 2024, SEXPAID has been slightly adjusted to help protect confidentiality. Please see the 2024 codebook for more information.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8204/vshow}{GSS Data Explorer page for \code{sexpaid}}.
}

\seealso{
Other High Risk Behaviors: 
\code{\link{sexornt1}}
}
\concept{High Risk Behaviors}
\keyword{variable}
