% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_139.R
\name{sexpromo}
\alias{sexpromo}
\title{Sex makes your promotion different}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexpromo
}
\details{
Question Do you think your being a (man/woman) makes your promotion opportunities better or worse?
}
\section{Values}{

\itemize{
\item \code{1} better
\item \code{2} worse
\item \code{3} no effect
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991, 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1304/vshow}{GSS Data Explorer page for \code{sexpromo}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab better \tab worse \tab no effect \tab don't know \tab no answer \tab Total \cr
   1991 \tab 605 \tab 265 \tab 115 \tab 505 \tab 11 \tab 16 \tab 1517 \cr
   1996 \tab 1446 \tab 416 \tab 306 \tab 609 \tab 53 \tab 74 \tab 2904 \cr
   Total \tab 2051 \tab 681 \tab 421 \tab 1114 \tab 64 \tab 90 \tab 4421 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{advances}},
\code{\link{futpromo}},
\code{\link{imppromo}},
\code{\link{jobcntrl}},
\code{\link{lottosay}},
\code{\link{numpromo}},
\code{\link{promoted}},
\code{\link{racpromo}},
\code{\link{wrkindep}}
}
\concept{Gender}
\concept{Work Organizations}
\keyword{variable}
