% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_426.R
\name{sexsex18}
\alias{sexsex18}
\title{Since age 18, have your sex partner's been...}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexsex18
}
\details{
Question Since the age of 18, have your sex partners been:
}
\section{Values}{

\itemize{
\item \code{1} only men
\item \code{2} some women, but mostly men
\item \code{3} equally men and women
\item \code{4} some men, but mostly women
\item \code{5} only women
\item \code{6} i have not had any sex partners
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3906/vshow}{GSS Data Explorer page for \code{sexsex18}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab only men \tab some women, but mostly men \tab equally men and women \tab some men, but mostly women \tab only women \tab i have not had any sex partners \tab don't know \tab no answer \tab Total \cr
   2008 \tab 239 \tab 890 \tab 36 \tab 7 \tab 13 \tab 754 \tab 31 \tab 3 \tab 50 \tab 2023 \cr
   2012 \tab 237 \tab 859 \tab 56 \tab 11 \tab 22 \tab 691 \tab 57 \tab - \tab 41 \tab 1974 \cr
   Total \tab 476 \tab 1749 \tab 92 \tab 18 \tab 35 \tab 1445 \tab 88 \tab 3 \tab 91 \tab 3997 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{toldsmsx}}
}
\concept{Sexual Orientation}
\keyword{variable}
