% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_427.R
\name{sexunion}
\alias{sexunion}
\title{Sex of r's spouse/civil union partner/domestic partner}
\source{
General Social Survey https://gss.norc.org
}
\description{
sexunion
}
\details{
Question What is the sex of your (spouse/civil union partner/domestic partner)?
}
\section{Values}{

\itemize{
\item \verb{[1]} male
\item \verb{[2]} female
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3912/vshow}{GSS Data Explorer page for \code{sexunion}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab male \tab female \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1946 \tab 25 \tab 19 \tab 1 \tab 32 \tab 2023 \cr
   Total \tab 1946 \tab 25 \tab 19 \tab 1 \tab 32 \tab 2023 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{adoptkid}},
\code{\link{attractd}},
\code{\link{evkid}},
\code{\link{kidlived}},
\code{\link{kidnow1}},
\code{\link{kidresp}},
\code{\link{marunion}},
\code{\link{relgendr}},
\code{\link{toldsxor}}
}
\concept{Sexual Orientation}
\keyword{variable}
