% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_644.R
\name{sharehhw}
\alias{sharehhw}
\title{Who in household does housework}
\source{
General Social Survey https://gss.norc.org
}
\description{
sharehhw
}
\details{
Question Which of the following best applies to the sharing of household work between you and your (spousepartfill)?
}
\section{Values}{

\itemize{
\item \verb{[1]} r does much more than fair share
\item \verb{[2]} r does a bit more than fair share
\item \verb{[3]} r does roughly fair share
\item \verb{[4]} r does a bit less than fair share
\item \verb{[5]} r does much less than fair share
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7776/vshow}{GSS Data Explorer page for \code{sharehhw}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab r does much more than fair share \tab r does a bit more than fair share \tab r does roughly fair share \tab r does a bit less than fair share \tab r does much less than fair share \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 3024 \tab 77 \tab 97 \tab 228 \tab 73 \tab 39 \tab 2 \tab 3 \tab 1 \tab 3544 \cr
   Total \tab 3024 \tab 77 \tab 97 \tab 228 \tab 73 \tab 39 \tab 2 \tab 3 \tab 1 \tab 3544 \cr
}
}

\keyword{variable}
