% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_177.R
\name{sharesep}
\alias{sharesep}
\title{Man and woman do things together-separate}
\source{
General Social Survey https://gss.norc.org
}
\description{
sharesep
}
\details{
Question Next, we have some questions about the kind of relationship you would like with a spouse or partner. It doesn't matter whether you are now married or living with someone. For each pair of items on the card, which type of relationship would you prefer? In pair A, would you prefer the type of relationship described in statement 1 or statement 2? How about in pair B? Pair C?
B.
}
\section{Values}{

\itemize{
\item \code{1} man woman do things together
\item \code{2} man woman do things separate
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1658/vshow}{GSS Data Explorer page for \code{sharesep}}.
}

\seealso{
Other Gender: 
\code{\link{emoteoth}},
\code{\link{fejobaff}},
\code{\link{hapboys}},
\code{\link{hapgirls}},
\code{\link{rhmewrk}},
\code{\link{sphmewrk}},
\code{\link{tradmod}},
\code{\link{willwed1}},
\code{\link{willwed2}}

Other Cohabitation: 
\code{\link{emoteoth}},
\code{\link{tradmod}}

Other Marriage: 
\code{\link{emoteoth}},
\code{\link{tradmod}},
\code{\link{willwed1}},
\code{\link{willwed2}}
}
\concept{Cohabitation}
\concept{Gender}
\concept{Marriage}
\keyword{variable}
