% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_486.R
\name{shop1}
\alias{shop1}
\title{Who in household shops for groceries}
\source{
General Social Survey https://gss.norc.org
}
\description{
shop1
}
\details{
Question In your household who does the following things...?
D. Shops for groceries?
}
\section{Values}{

\itemize{
\item \code{1} always me
\item \code{2} usually me
\item \code{3} about equal or both together
\item \code{4} usually my spouse/partner
\item \code{5} always my spouse/partner
\item \code{6} is done by a third person
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2012 \tab A/B/- \tab partial \cr
   2022 \tab A/-/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4432/vshow}{GSS Data Explorer page for \code{shop1}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{caresik1}},
\code{\link{clean1}},
\code{\link{cooking1}},
\code{\link{deckids}},
\code{\link{hhwkdis}},
\code{\link{hhwkfair}},
\code{\link{rhhwork}},
\code{\link{spborn}},
\code{\link{sphhwork}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
