% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_485.R
\name{shopfood}
\alias{shopfood}
\title{Who in household shops for groceries}
\source{
General Social Survey https://gss.norc.org
}
\description{
shopfood
}
\details{
Question In your household who does the following things . . .
D. Shops for groceries.
}
\section{Values}{

\itemize{
\item \code{1} always the woman
\item \code{2} usually the woman
\item \code{3} about equal or both together
\item \code{4} usually the man
\item \code{5} always the man
\item \code{6} is done by a third person
\item \code{7} not married
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4421/vshow}{GSS Data Explorer page for \code{shopfood}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{dinner}},
\code{\link{laundry1}},
\code{\link{mehhwork}},
\code{\link{mekdcare}},
\code{\link{repairs1}},
\code{\link{spwkbaby}},
\code{\link{spwkgrwn}},
\code{\link{spwknokd}},
\code{\link{spwrksch}}

Other Household: 
\code{\link{dinner}},
\code{\link{laundry1}},
\code{\link{repairs1}}
}
\concept{Household}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
