% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_124.R
\name{shopnum}
\alias{shopnum}
\title{How many churches did you visit}
\source{
General Social Survey https://gss.norc.org
}
\description{
shopnum
}
\details{
Question When you most recently looked for a church or synagogue, how many different places of worship did you visit?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1157/vshow}{GSS Data Explorer page for \code{shopnum}}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 15 \tab 2 \tab 20 \tab 22 \tab 3 \tab 4 \tab 5 \tab 6 \tab 9 \tab don't know \tab no answer \tab Total \cr
   1998 \tab 2258 \tab 2 \tab 66 \tab 6 \tab 2 \tab 171 \tab 2 \tab 1 \tab 152 \tab 69 \tab 48 \tab 18 \tab 1 \tab 5 \tab 31 \tab 2832 \cr
   Total \tab 2258 \tab 2 \tab 66 \tab 6 \tab 2 \tab 171 \tab 2 \tab 1 \tab 152 \tab 69 \tab 48 \tab 18 \tab 1 \tab 5 \tab 31 \tab 2832 \cr
}
}

\seealso{
Other Religion: 
\code{\link{chrissch}},
\code{\link{denomsch}},
\code{\link{implives}},
\code{\link{othchrch}},
\code{\link{othsch}},
\code{\link{privsch}},
\code{\link{relsch}},
\code{\link{shopmove}},
\code{\link{shoprel}}
}
\concept{Religion}
\keyword{variable}
