% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_456.R
\name{sibmost}
\alias{sibmost}
\title{Do r have the most contact with sister or brother}
\source{
General Social Survey https://gss.norc.org
}
\description{
sibmost
}
\details{
Question Of your adult brothers and sisters, with whom do you have the most contact?
}
\section{Values}{

\itemize{
\item \verb{[1]} with a brother
\item \verb{[2]} with a sister
\item \verb{[3]} i have no contact with any adult brother or sister
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4185/vshow}{GSS Data Explorer page for \code{sibmost}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab with a brother \tab with a sister \tab i have no contact with any adult brother or sister \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1745 \tab 414 \tab 537 \tab 58 \tab 4 \tab 7 \tab 2765 \cr
   Total \tab 1745 \tab 414 \tab 537 \tab 58 \tab 4 \tab 7 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{kidmost}},
\code{\link{kidnum}},
\code{\link{kidvisit}},
\code{\link{livecom}},
\code{\link{livehome}},
\code{\link{matime1}},
\code{\link{sibcall}},
\code{\link{sibnum}},
\code{\link{sibvisit}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
