% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_555.R
\name{siborder}
\alias{siborder}
\title{Birth order}
\source{
General Social Survey https://gss.norc.org
}
\description{
siborder
}
\details{
Question How many of your (number) brothers and sisters were born before you were?
}
\section{Values}{

\itemize{
\item \code{95} only child
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5082/vshow}{GSS Data Explorer page for \code{siborder}}.

Counts by year:\tabular{lllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 16 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab only child \tab don't know \tab no answer \tab Total \cr
   1990 \tab 444 \tab 231 \tab 244 \tab 5 \tab 2 \tab 1 \tab 148 \tab 86 \tab 58 \tab 36 \tab 18 \tab 18 \tab 11 \tab 6 \tab 45 \tab 1 \tab 18 \tab 1372 \cr
   Total \tab 444 \tab 231 \tab 244 \tab 5 \tab 2 \tab 1 \tab 148 \tab 86 \tab 58 \tab 36 \tab 18 \tab 18 \tab 11 \tab 6 \tab 45 \tab 1 \tab 18 \tab 1372 \cr
}
}

\concept{Birth Order}
\keyword{variable}
