% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_555.R
\name{siborder}
\alias{siborder}
\title{Birth order}
\source{
General Social Survey https://gss.norc.org
}
\description{
siborder
}
\details{
Question How many of your (number) brothers and sisters were born before you were?
}
\section{Values}{

\itemize{
\item \code{95} only child
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5082/vshow}{GSS Data Explorer page for \code{siborder}}.
}

\concept{Birth Order}
\keyword{variable}
