% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_456.R
\name{sibvisit}
\alias{sibvisit}
\title{How often does r visit this sister or brother}
\source{
General Social Survey https://gss.norc.org
}
\description{
sibvisit
}
\details{
Question How often do you see or visit this brother or sister?
}
\section{Values}{

\itemize{
\item \verb{[1]} he or she lives in the same household as i do
\item \verb{[2]} daily
\item \verb{[3]} at least several times a week
\item \verb{[4]} at least once a week
\item \verb{[5]} at least once a month
\item \verb{[6]} several times a year
\item \verb{[7]} less often
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4186/vshow}{GSS Data Explorer page for \code{sibvisit}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab he or she lives in the same household as i do \tab daily \tab at least several times a week \tab at least once a week \tab at least once a month \tab several times a year \tab less often \tab no answer \tab Total \cr
   2002 \tab 1807 \tab 25 \tab 67 \tab 109 \tab 160 \tab 172 \tab 226 \tab 192 \tab 7 \tab 2765 \cr
   Total \tab 1807 \tab 25 \tab 67 \tab 109 \tab 160 \tab 172 \tab 226 \tab 192 \tab 7 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{kidmost}},
\code{\link{kidnum}},
\code{\link{kidvisit}},
\code{\link{livecom}},
\code{\link{livehome}},
\code{\link{matime1}},
\code{\link{sibcall}},
\code{\link{sibmost}},
\code{\link{sibnum}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
