% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_456.R
\name{sibvisit}
\alias{sibvisit}
\title{How often does r visit this sister or brother}
\source{
General Social Survey https://gss.norc.org
}
\description{
sibvisit
}
\details{
Question How often do you see or visit this brother or sister?
}
\section{Values}{

\itemize{
\item \code{1} he or she lives in the same household as i do
\item \code{2} daily
\item \code{3} at least several times a week
\item \code{4} at least once a week
\item \code{5} at least once a month
\item \code{6} several times a year
\item \code{7} less often
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4186/vshow}{GSS Data Explorer page for \code{sibvisit}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{kidmost}},
\code{\link{kidnum}},
\code{\link{kidvisit}},
\code{\link{livecom}},
\code{\link{livehome}},
\code{\link{matime1}},
\code{\link{sibcall}},
\code{\link{sibmost}},
\code{\link{sibnum}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
