% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_459.R
\name{sick1a}
\alias{sick1a}
\title{Who does r turn to first for help when sick}
\source{
General Social Survey https://gss.norc.org
}
\description{
sick1a
}
\details{
Question Now we would like to ask you how you would get help in situations that anyone could find herself or himself in. First, suppose you had the 'flu and had to stay in bed for a few days and needed help around the house, with shopping and so on.
A. Who would you turn to first for help?
}
\section{Values}{

\itemize{
\item \verb{[1]} husband, wife, partner
\item \verb{[2]} mother
\item \verb{[3]} father
\item \verb{[4]} daughter
\item \verb{[5]} daughter-in-law
\item \verb{[6]} son
\item \verb{[7]} son-in-law
\item \verb{[8]} sister
\item \verb{[9]} brother
\item \verb{[10]} other blood relative
\item \verb{[11]} other in-law relative
\item \verb{[12]} close friend
\item \verb{[13]} neighbor
\item \verb{[14]} someone you work with
\item \verb{[15]} someone at a social services agency
\item \verb{[16]} someone you pay to help
\item \verb{[17]} someone else
\item \verb{[18]} no one
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4214/vshow}{GSS Data Explorer page for \code{sick1a}}.

Counts by year:\tabular{llllllllllllllllllllll}{
   year \tab iap \tab husband, wife, partner \tab mother \tab father \tab daughter \tab daughter-in-law \tab son \tab sister \tab brother \tab other blood relative \tab other in-law relative \tab close friend \tab neighbor \tab someone you work with \tab someone at a social services agency \tab someone you pay to help \tab someone else \tab no one \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1616 \tab 547 \tab 137 \tab 26 \tab 98 \tab 7 \tab 33 \tab 51 \tab 27 \tab 17 \tab 7 \tab 113 \tab 25 \tab 6 \tab 3 \tab 2 \tab 14 \tab 29 \tab 1 \tab 6 \tab 2765 \cr
   Total \tab 1616 \tab 547 \tab 137 \tab 26 \tab 98 \tab 7 \tab 33 \tab 51 \tab 27 \tab 17 \tab 7 \tab 113 \tab 25 \tab 6 \tab 3 \tab 2 \tab 14 \tab 29 \tab 1 \tab 6 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{borrow1a}},
\code{\link{borrow2a}},
\code{\link{down1a}},
\code{\link{down2a}},
\code{\link{grpchrty}},
\code{\link{grpnei}},
\code{\link{grpoth}},
\code{\link{helphwrk}},
\code{\link{sick2a}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
