% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_459.R
\name{sick2a}
\alias{sick2a}
\title{Who does r turn to second for help when sick}
\source{
General Social Survey https://gss.norc.org
}
\description{
sick2a
}
\details{
Question Now we would like to ask you how you would get help in situations that anyone could find herself or himself in. First, suppose you had the 'flu and had to stay in bed for a few days and needed help around the house, with shopping and so on.
B. And who would you turn to second if you had the flu and needed help around the house?
}
\section{Values}{

\itemize{
\item \code{1} husband, wife, partner
\item \code{2} mother
\item \code{3} father
\item \code{4} daughter
\item \code{5} daughter-in-law
\item \code{6} son
\item \code{7} son-in-law
\item \code{8} sister
\item \code{9} brother
\item \code{10} other blood relative
\item \code{11} other in-law relative
\item \code{12} close friend
\item \code{13} neighbor
\item \code{14} someone you work with
\item \code{15} someone at a social services agency
\item \code{16} someone you pay to help
\item \code{17} other
\item \code{18} no one
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4215/vshow}{GSS Data Explorer page for \code{sick2a}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{borrow1a}},
\code{\link{borrow2a}},
\code{\link{down1a}},
\code{\link{down2a}},
\code{\link{grpchrty}},
\code{\link{grpnei}},
\code{\link{grpoth}},
\code{\link{helphwrk}},
\code{\link{sick1a}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
