% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_451.R
\name{sisnum}
\alias{sisnum}
\title{Number of adult sister's}
\source{
General Social Survey https://gss.norc.org
}
\description{
sisnum
}
\details{
Question How many sisters age 18 or older do you have? (We mean sisters who are still alive; please include step-sisters, half-sisters, and adopted sisters.)
}
\section{Values}{

\itemize{
\item \code{0} none
\item \code{1} one
\item \code{2} two
\item \code{3} three
\item \code{4} four
\item \code{5} five +
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1986 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4135/vshow}{GSS Data Explorer page for \code{sisnum}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{bronum}},
\code{\link{brotime}},
\code{\link{brovisit}},
\code{\link{pacall}},
\code{\link{patime}},
\code{\link{pavisit}},
\code{\link{siscall}},
\code{\link{sistime}},
\code{\link{sisvisit}}

Other Family: 
\code{\link{bronum}},
\code{\link{brotime}},
\code{\link{brovisit}},
\code{\link{pacall}},
\code{\link{patime}},
\code{\link{pavisit}},
\code{\link{siscall}},
\code{\link{sistime}},
\code{\link{sisvisit}}
}
\concept{Family}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
