% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_239.R
\name{siteyrs}
\alias{siteyrs}
\title{How long have worked for those sites}
\source{
General Social Survey https://gss.norc.org
}
\description{
siteyrs
}
\details{
Question B. How long have you worked full time for those work sites? How many years?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2148/vshow}{GSS Data Explorer page for \code{siteyrs}}.
}

\seealso{
Other Jobs: 
\code{\link{formltrn}},
\code{\link{genderwk}},
\code{\link{locyrs}},
\code{\link{orgsize}},
\code{\link{othyrs}},
\code{\link{otjtrain}},
\code{\link{partorg}},
\code{\link{samework}},
\code{\link{wrkoth}}

Other Job: 
\code{\link{formltrn}},
\code{\link{genderwk}},
\code{\link{locyrs}},
\code{\link{orgsize}},
\code{\link{othyrs}},
\code{\link{otjtrain}},
\code{\link{partorg}},
\code{\link{samework}},
\code{\link{wrkoth}}
}
\concept{Job}
\concept{Jobs}
\keyword{variable}
