% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_631.R
\name{skipmeal}
\alias{skipmeal}
\title{How often r's household skip a meal}
\source{
General Social Survey https://gss.norc.org
}
\description{
skipmeal
}
\details{
Question How often do you or other members of your household skip a meal because there is not enough money for food?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} less than once a month
\item \code{3} once a month
\item \code{4} several times a month
\item \code{5} once a week
\item \code{6} several times a week
\item \code{7} every day
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7519/vshow}{GSS Data Explorer page for \code{skipmeal}}.
}

\keyword{variable}
