% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_508.R
\name{skltrain}
\alias{skltrain}
\title{R had job skill training in thepast 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
skltrain
}
\details{
Question In addition to your main job, do you do any other work for additional income?
IF YES F. Over the past 12 months, have you had any training to improve your job skills?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D/- \tab partial \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4625/vshow}{GSS Data Explorer page for \code{skltrain}}.
}

\seealso{
Other Work Orientation: 
\code{\link{nounemp3}},
\code{\link{nounemp4}},
\code{\link{otherwrk}},
\code{\link{spwrkgvt}}
}
\concept{Work Orientation}
\keyword{variable}
