% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_391.R
\name{slfmangd}
\alias{slfmangd}
\title{R involved in a self-managed team}
\source{
General Social Survey https://gss.norc.org
}
\description{
slfmangd
}
\details{
Question A. Are you currently involved in a self-managed team?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3558/vshow}{GSS Data Explorer page for \code{slfmangd}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2006 \tab 2778 \tab 568 \tab 1118 \tab 25 \tab 21 \tab - \tab 4510 \cr
   2018 \tab 900 \tab 447 \tab 962 \tab 7 \tab 32 \tab - \tab 2348 \cr
   2022 \tab 2570 \tab 341 \tab 569 \tab 49 \tab 8 \tab 7 \tab 3544 \cr
   Total \tab 6248 \tab 1356 \tab 2649 \tab 81 \tab 61 \tab 7 \tab 10402 \cr
}
}

\seealso{
Other Shared Capitalism: 
\code{\link{compwage}},
\code{\link{empinput}},
\code{\link{emptrain}},
\code{\link{esop}},
\code{\link{extr2005}},
\code{\link{numorg}},
\code{\link{ops2005}},
\code{\link{talkteam}},
\code{\link{wealth}}
}
\concept{Shared Capitalism}
\keyword{variable}
