% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_391.R
\name{slfmangd}
\alias{slfmangd}
\title{R involved in a self-managed team}
\source{
General Social Survey https://gss.norc.org
}
\description{
slfmangd
}
\details{
Question A. Are you currently involved in a self-managed team?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3558/vshow}{GSS Data Explorer page for \code{slfmangd}}.
}

\seealso{
Other Shared Capitalism: 
\code{\link{compwage}},
\code{\link{empinput}},
\code{\link{emptrain}},
\code{\link{esop}},
\code{\link{extr2005}},
\code{\link{numorg}},
\code{\link{ops2005}},
\code{\link{talkteam}},
\code{\link{wealth}}

Other Work: 
\code{\link{compwage}},
\code{\link{empinput}},
\code{\link{emptrain}},
\code{\link{esop}},
\code{\link{extr2005}},
\code{\link{numorg}},
\code{\link{ops2005}},
\code{\link{talkteam}},
\code{\link{wealth}}
}
\concept{Shared Capitalism}
\concept{Work}
\keyword{variable}
