% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_312.R
\name{slpprblm}
\alias{slpprblm}
\title{Trouble sleeping last 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
slpprblm
}
\details{
Question B. During the past 12 months, how often have you had trouble going to sleep or staying asleep?
}
\section{Values}{

\itemize{
\item \code{1} often
\item \code{2} sometimes
\item \code{3} rarely
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2837/vshow}{GSS Data Explorer page for \code{slpprblm}}.
}

\seealso{
Other Quality of Working Life: 
\code{\link{arthrtis}},
\code{\link{depress}},
\code{\link{diabetes}},
\code{\link{hyperten}},
\code{\link{knowschd}},
\code{\link{satjob1}},
\code{\link{stress12}},
\code{\link{usetech}},
\code{\link{weight}}
}
\concept{Quality of Working Life}
\keyword{variable}
