% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_312.R
\name{slpprblm}
\alias{slpprblm}
\title{Trouble sleeping last 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
slpprblm
}
\details{
Question B. During the past 12 months, how often have you had trouble going to sleep or staying asleep?
}
\section{Values}{

\itemize{
\item \code{1} often
\item \code{2} sometimes
\item \code{3} rarely
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2837/vshow}{GSS Data Explorer page for \code{slpprblm}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab often \tab sometimes \tab rarely \tab never \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2010 \tab 857 \tab 258 \tab 384 \tab 315 \tab 206 \tab 1 \tab 23 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 274 \tab 432 \tab 324 \tab 216 \tab - \tab 4 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 316 \tab 505 \tab 395 \tab 201 \tab 1 \tab 30 \tab - \tab 2348 \cr
   2022 \tab 1556 \tab 484 \tab 716 \tab 534 \tab 226 \tab 3 \tab 10 \tab 15 \tab 3544 \cr
   Total \tab 4601 \tab 1332 \tab 2037 \tab 1568 \tab 849 \tab 5 \tab 67 \tab 15 \tab 10474 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{arthrtis}},
\code{\link{depress}},
\code{\link{diabetes}},
\code{\link{hyperten}},
\code{\link{knowschd}},
\code{\link{satjob1}},
\code{\link{stress12}},
\code{\link{usetech}},
\code{\link{weight}}
}
\concept{Quality of Working Life}
\keyword{variable}
