% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_60.R
\name{smokecig}
\alias{smokecig}
\title{Does r smoke cigarettes}
\source{
General Social Survey https://gss.norc.org
}
\description{
smokecig
}
\details{
Question Do you smoke?
B. Do you smoke cigarettes?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977, 1978, 1980, 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/534/vshow}{GSS Data Explorer page for \code{smokecig}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1977 \tab 878 \tab 584 \tab 54 \tab 14 \tab 1530 \cr
   1978 \tab 927 \tab 545 \tab 53 \tab 7 \tab 1532 \cr
   1980 \tab 866 \tab 551 \tab 43 \tab 8 \tab 1468 \cr
   1983 \tab 1006 \tab 561 \tab 28 \tab 4 \tab 1599 \cr
   1984 \tab 924 \tab 514 \tab 27 \tab 8 \tab 1473 \cr
   Total \tab 4601 \tab 2755 \tab 205 \tab 41 \tab 7602 \cr
}
}

\seealso{
Other Core: 
\code{\link{anomia1}},
\code{\link{anomia2}},
\code{\link{anomia3}},
\code{\link{anomia4}},
\code{\link{anomia5}},
\code{\link{anomia6}},
\code{\link{cigweek}},
\code{\link{evsmoke}},
\code{\link{quitsmk}}
}
\concept{Core}
\keyword{variable}
