% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_650.R
\name{smokeday}
\alias{smokeday}
\title{How many cigs a day}
\source{
General Social Survey https://gss.norc.org
}
\description{
smokeday
}
\details{
Question Do you smoke cigarettes, and if so about how many cigarettes a day?
}
\section{Values}{

\itemize{
\item \verb{[1]} do not smoke and never did
\item \verb{[2]} do not smoke now but smoked in the past
\item \verb{[3]} smoke 1-5 cigarettes per day
\item \verb{[4]} smoke 6-10 cigarettes per day
\item \verb{[5]} smoke 11-20 cigarettes per day
\item \verb{[6]} smoke 21-40 cigarettes per day
\item \verb{[7]} smoke more than 40 cigarettes per day
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7834/vshow}{GSS Data Explorer page for \code{smokeday}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab do not smoke and never did \tab do not smoke now but smoked in the past \tab smoke 1-5 cigarettes per day \tab smoke 6-10 cigarettes per day \tab smoke 11-20 cigarettes per day \tab smoke 21-40 cigarettes per day \tab smoke more than 40 cigarettes per day \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2382 \tab 665 \tab 279 \tab 53 \tab 53 \tab 56 \tab 17 \tab 5 \tab 25 \tab 3 \tab 6 \tab 3544 \cr
   Total \tab 2382 \tab 665 \tab 279 \tab 53 \tab 53 \tab 56 \tab 17 \tab 5 \tab 25 \tab 3 \tab 6 \tab 3544 \cr
}
}

\keyword{variable}
