% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_650.R
\name{smokeday}
\alias{smokeday}
\title{How many cigs a day}
\source{
General Social Survey https://gss.norc.org
}
\description{
smokeday
}
\details{
Question Do you smoke cigarettes, and if so about how many cigarettes a day?
}
\section{Values}{

\itemize{
\item \code{1} do not smoke and never did
\item \code{2} do not smoke now but smoked in the past
\item \code{3} smoke 1-5 cigarettes per day
\item \code{4} smoke 6-10 cigarettes per day
\item \code{5} smoke 11-20 cigarettes per day
\item \code{6} smoke 21-40 cigarettes per day
\item \code{7} smoke more than 40 cigarettes per day
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7834/vshow}{GSS Data Explorer page for \code{smokeday}}.
}

\keyword{variable}
