% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_288.R
\name{snsmfrst}
\alias{snsmfrst}
\title{Which social network did you join first}
\source{
General Social Survey https://gss.norc.org
}
\description{
snsmfrst
}
\details{
Question Of the social networking/social media sites that you mentioned, which did you join or start using first? 
SOCIAL NETWORKING/SOCIAL MEDIA SITE:
}
\section{Values}{

\itemize{
\item \verb{[1]} twitter
\item \verb{[2]} facebook
\item \verb{[3]} instagram
\item \verb{[4]} linkedin
\item \verb{[5]} snapchat
\item \verb{[6]} tumblr
\item \verb{[7]} whatsapp
\item \verb{[8]} google+
\item \verb{[9]} pinterest
\item \verb{[10]} flickr
\item \verb{[11]} vine
\item \verb{[12]} classmates
\item \verb{[13]} youtube
\item \verb{[14]} myspace
\item \verb{[15]} reddit
\item \verb{[16]} viber
\item \verb{[17]} line
\item \verb{[18]} plenty of fish
\item \verb{[19]} kik
\item \verb{[20]} ravelry
\item \verb{[22]} wechat
\item \verb{[23]} imo
\item \verb{[24]} online dating sites
\item \verb{[25]} spotify
\item \verb{[26]} groupme
\item \verb{[27]} tinder
\item \verb{[28]} hangouts
\item \verb{[29]} researchgate
\item \verb{[30]} online gaming
\item \verb{[31]} yahoo
\item \verb{[32]} other social networks
\item \verb{[33]} not a social network
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6286/vshow}{GSS Data Explorer page for \code{snsmfrst}}.

Counts by year:\tabular{lllllllllllllllllllllll}{
   year \tab iap \tab twitter \tab facebook \tab instagram \tab linkedin \tab snapchat \tab tumblr \tab whatsapp \tab google+ \tab pinterest \tab flickr \tab classmates \tab youtube \tab plenty of fish \tab kik \tab online dating sites \tab online gaming \tab yahoo \tab other social networks \tab not a social network \tab no answer \tab Total \cr
   2016 \tab 1871 \tab 29 \tab 688 \tab 21 \tab 58 \tab 15 \tab 4 \tab 6 \tab 90 \tab 31 \tab 4 \tab 16 \tab 7 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 6 \tab 10 \tab 2867 \cr
   Total \tab 1871 \tab 29 \tab 688 \tab 21 \tab 58 \tab 15 \tab 4 \tab 6 \tab 90 \tab 31 \tab 4 \tab 16 \tab 7 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 6 \tab 10 \tab 2867 \cr
}
}

\seealso{
Other Social Media Usage: 
\code{\link{intwkdyh}},
\code{\link{intwkdym}},
\code{\link{intwkenh}},
\code{\link{intwkenm}}
}
\concept{Social Media Usage}
\keyword{variable}
