% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_287.R
\name{snsmnew}
\alias{snsmnew}
\title{Which social network did you join most recently}
\source{
General Social Survey https://gss.norc.org
}
\description{
snsmnew
}
\details{
Question Of the social networking/social media sites that you mentioned, which did you join or start using most recently?
SOCIAL NETWORKING/SOCIAL MEDIA SITE:
}
\section{Values}{

\itemize{
\item \code{1} twitter
\item \code{2} facebook
\item \code{3} instagram
\item \code{4} linkedin
\item \code{5} snapchat
\item \code{6} tumblr
\item \code{7} whatsapp
\item \code{8} google+
\item \code{9} pinterest
\item \code{10} flickr
\item \code{11} vine
\item \code{12} classmates
\item \code{13} youtube
\item \code{14} myspace
\item \code{15} reddit
\item \code{16} viber
\item \code{17} line
\item \code{18} plenty of fish
\item \code{19} kik
\item \code{20} ravelry
\item \code{22} wechat
\item \code{23} imo
\item \code{24} online dating sites
\item \code{25} spotify
\item \code{26} groupme
\item \code{27} tinder
\item \code{28} hangouts
\item \code{29} researchgate
\item \code{30} online gaming
\item \code{31} yahoo
\item \code{32} other social networks
\item \code{33} not a social network
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6285/vshow}{GSS Data Explorer page for \code{snsmnew}}.
}

\seealso{
Other Social Media Usage: 
\code{\link{clssmtes}},
\code{\link{facebook}},
\code{\link{flickr}},
\code{\link{googlesn}},
\code{\link{instagrm}},
\code{\link{intmbile}},
\code{\link{intrecnt}},
\code{\link{intuse}},
\code{\link{intwkdyh}},
\code{\link{intwkdym}},
\code{\link{intwkenh}},
\code{\link{intwkenm}},
\code{\link{linkedin}},
\code{\link{pinterst}},
\code{\link{snapchat}},
\code{\link{snsmfrst}},
\code{\link{snsmot2a}},
\code{\link{snsmot2b}},
\code{\link{snsmot2c}},
\code{\link{snsmoth1}},
\code{\link{tumblr}},
\code{\link{twitter}},
\code{\link{vine}},
\code{\link{whatsapp}}
}
\concept{Social Media Usage}
\keyword{variable}
