% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_287.R
\name{snsmot2c}
\alias{snsmot2c}
\title{Third other social network}
\source{
General Social Survey https://gss.norc.org
}
\description{
snsmot2c
}
\details{
Question THIRD MENTION:
}
\section{Values}{

\itemize{
\item \code{1} youtube
\item \code{2} myspace
\item \code{3} reddit
\item \code{4} viber
\item \code{5} line
\item \code{6} plenty of fish
\item \code{7} kik
\item \code{8} ravelry
\item \code{9} wechat
\item \code{11} imo
\item \code{12} online dating sites
\item \code{13} spotify
\item \code{14} groupme
\item \code{15} tinder
\item \code{16} hangouts
\item \code{17} researchgate
\item \code{18} online gaming
\item \code{19} yahoo
\item \code{30} other social networks
\item \code{40} not a social network
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6284/vshow}{GSS Data Explorer page for \code{snsmot2c}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab youtube \tab viber \tab kik \tab hangouts \tab other social networks \tab not a social network \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2845 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 11 \tab 2867 \cr
   Total \tab 2845 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 4 \tab 1 \tab 11 \tab 2867 \cr
}
}

\seealso{
Other Social Media Usage: 
\code{\link{clssmtes}},
\code{\link{flickr}},
\code{\link{pinterst}},
\code{\link{snsmnew}},
\code{\link{snsmot2a}},
\code{\link{snsmot2b}},
\code{\link{snsmoth1}},
\code{\link{vine}}
}
\concept{Social Media Usage}
\keyword{variable}
