% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_287.R
\name{snsmyear}
\alias{snsmyear}
\title{Year first joined sns}
\source{
General Social Survey https://gss.norc.org
}
\description{
snsmyear
}
\details{
Question In what year did you start using the internet or web-enabled applications/APPS?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6648/vshow}{GSS Data Explorer page for \code{snsmyear}}.
}

\keyword{variable}
