% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_296.R
\name{socacts}
\alias{socacts}
\title{Physical and emotion social activities past 4}
\source{
General Social Survey https://gss.norc.org
}
\description{
socacts
}
\details{
Question During the past 4 weeks, how much of the time has your physical health or emotional problems interfered with your social activities (like visiting friends, relatives, etc.)?
}
\section{Values}{

\itemize{
\item \verb{[1]} all of the time
\item \verb{[2]} most of the time
\item \verb{[3]} a good bit of the time
\item \verb{[4]} some of the time
\item \verb{[5]} a little bit of the time
\item \verb{[6]} non of the time
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2687/vshow}{GSS Data Explorer page for \code{socacts}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab all of the time \tab most of the time \tab a good bit of the time \tab some of the time \tab a little bit of the time \tab non of the time \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 40 \tab 49 \tab 52 \tab 134 \tab 196 \tab 936 \tab 12 \tab 2817 \cr
   Total \tab 1398 \tab 40 \tab 49 \tab 52 \tab 134 \tab 196 \tab 936 \tab 12 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{downblue}},
\code{\link{energy}},
\code{\link{pain}},
\code{\link{peaceful}},
\code{\link{treat1}},
\code{\link{treat2}},
\code{\link{treat3}},
\code{\link{treat4}},
\code{\link{treat5}}
}
\concept{Health Status}
\keyword{variable}
