% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_126.R
\name{socdif2}
\alias{socdif2}
\title{Paid what you need}
\source{
General Social Survey https://gss.norc.org
}
\description{
socdif2
}
\details{
Question Here are different opinions about social differences in this country. Please tell me for each one whether you strongly agree, somewhat agree, somewhat disagree, or strongly disagree.
B. Personal income should not be determined solely by one's work.  Rather, everybody should get what he/she needs to provide a decent life to his/her family.
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree
\item \code{3} disagree
\item \code{4} disagree strongly
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1175/vshow}{GSS Data Explorer page for \code{socdif2}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab disagree \tab disagree strongly \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 86 \tab 459 \tab 615 \tab 159 \tab 47 \tab 6 \tab 1372 \cr
   Total \tab 0 \tab 86 \tab 459 \tab 615 \tab 159 \tab 47 \tab 6 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{socdif1}},
\code{\link{socdif3}},
\code{\link{socdif4}},
\code{\link{whypoor1}},
\code{\link{whypoor2}},
\code{\link{whypoor3}},
\code{\link{whypoor4}},
\code{\link{wlthjews}},
\code{\link{wlthwhts}}
}
\concept{Intergroup Relations}
\keyword{variable}
