% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_58.R
\name{socfrend}
\alias{socfrend}
\title{Spend evening with friends}
\source{
General Social Survey https://gss.norc.org
}
\description{
socfrend
}
\details{
Question Would you use this card and tell me which answer comes closest to how often you do the following things…
C. Spend a social evening with friends who live outside the neighborhood?
}
\section{Values}{

\itemize{
\item \code{1} almost daily
\item \code{2} once or twice a week
\item \code{3} several times a month
\item \code{4} about once a month
\item \code{5} several times a year
\item \code{6} about once a year
\item \code{7} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1974 \tab -/- \tab full \cr
   1975 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1978 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1988 \tab A/B \tab partial \cr
   1989 \tab A/B \tab partial \cr
   1990 \tab A/B \tab partial \cr
   1991 \tab A/B \tab partial \cr
   1993 \tab A/B \tab partial \cr
   1994 \tab A/B \tab partial \cr
   1996 \tab A/B \tab partial \cr
   1998 \tab A/B \tab partial \cr
   2000 \tab A/B \tab partial \cr
   2002 \tab A/B \tab partial \cr
   2004 \tab A/B \tab partial \cr
   2006 \tab A/B \tab partial \cr
   2008 \tab A/B \tab partial \cr
   2010 \tab A/B \tab partial \cr
   2012 \tab A/B \tab partial \cr
   2014 \tab A/B \tab partial \cr
   2016 \tab A/B \tab partial \cr
   2018 \tab A/B \tab partial \cr
   2021 \tab A/B \tab partial \cr
   2022 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/523/vshow}{GSS Data Explorer page for \code{socfrend}}.
}

\seealso{
Other Core: 
\code{\link{chldsex}},
\code{\link{chldsex1}},
\code{\link{helpoth}},
\code{\link{socbar}},
\code{\link{socommun}},
\code{\link{socrel}},
\code{\link{thnkself}},
\code{\link{workhard}},
\code{\link{youngen}}

Other Sociability: 
\code{\link{socbar}},
\code{\link{socommun}},
\code{\link{socrel}}
}
\concept{Core}
\concept{Sociability}
\keyword{variable}
