% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_535.R
\name{socid1}
\alias{socid1}
\title{Most important to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
socid1
}
\details{
Question We are all part of different groups. Some are more important to us than others when we think of ourselves. in general, which in the following list is most important to you in describing who you are?  and the second most important? and the third most important?
A. Most Important
}
\section{Values}{

\itemize{
\item \code{1} current occupation
\item \code{2} race ethnic background
\item \code{3} gender
\item \code{4} age group
\item \code{5} religion
\item \code{6} political party
\item \code{7} nationality
\item \code{8} family or marital status
\item \code{9} social class
\item \code{10} region
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4878/vshow}{GSS Data Explorer page for \code{socid1}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{amancstr}},
\code{\link{citworld}},
\code{\link{decsorgs}},
\code{\link{freetrde}},
\code{\link{intlincs}},
\code{\link{lessprd}},
\code{\link{powrorgs}},
\code{\link{socid2}},
\code{\link{socid3}}

Other Occupation: 
\code{\link{socid2}},
\code{\link{socid3}}
}
\concept{ISSP National Identity}
\concept{Occupation}
\keyword{variable}
