% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_249.R
\name{socsecfx}
\alias{socsecfx}
\title{How serious is social securities problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
socsecfx
}
\details{
Question Please tell me which of the following statements comes closest to your opinion about the Social Security program …
}
\section{Values}{

\itemize{
\item \verb{[1]} no problems
\item \verb{[2]} minor probs
\item \verb{[3]} serious prob
\item \verb{[4]} be replaced
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2243/vshow}{GSS Data Explorer page for \code{socsecfx}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no problems \tab minor probs \tab serious prob \tab be replaced \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 64 \tab 304 \tab 714 \tab 218 \tab 80 \tab 7 \tab 2832 \cr
   Total \tab 1445 \tab 64 \tab 304 \tab 714 \tab 218 \tab 80 \tab 7 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{prozkid3}},
\code{\link{socsecnu}},
\code{\link{socsecrt}},
\code{\link{termecon}},
\code{\link{termemot}},
\code{\link{terminal}},
\code{\link{termpain}},
\code{\link{termpay}},
\code{\link{termrel}}
}
\concept{Mental Health}
\keyword{variable}
