% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_249.R
\name{socsecfx}
\alias{socsecfx}
\title{How serious is social securities problem}
\source{
General Social Survey https://gss.norc.org
}
\description{
socsecfx
}
\details{
Question Please tell me which of the following statements comes closest to your opinion about the Social Security program …
}
\section{Values}{

\itemize{
\item \code{1} no problems
\item \code{2} minor probs
\item \code{3} serious prob
\item \code{4} be replaced
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2243/vshow}{GSS Data Explorer page for \code{socsecfx}}.
}

\seealso{
Other Mental Health: 
\code{\link{prozkid3}},
\code{\link{socsecnu}},
\code{\link{socsecrt}},
\code{\link{termecon}},
\code{\link{termemot}},
\code{\link{terminal}},
\code{\link{termpain}},
\code{\link{termpay}},
\code{\link{termrel}}
}
\concept{Mental Health}
\keyword{variable}
