% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_249.R
\name{socsecnu}
\alias{socsecnu}
\title{Social security phase out or continue}
\source{
General Social Survey https://gss.norc.org
}
\description{
socsecnu
}
\details{
Question Do you think it would be better to phase out Social Security and require people to take what they and their employers now pay into Social Security and invest it in mandatory private retirement accounts, or do you think Social Security should continue much as it is now?
}
\section{Values}{

\itemize{
\item \verb{[1]} phase out social security+
\item \verb{[2]} continue as is
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2244/vshow}{GSS Data Explorer page for \code{socsecnu}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab phase out social security+ \tab continue as is \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 448 \tab 755 \tab 173 \tab 11 \tab 2832 \cr
   Total \tab 1445 \tab 448 \tab 755 \tab 173 \tab 11 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{prozkid3}},
\code{\link{socsecfx}},
\code{\link{socsecrt}},
\code{\link{termecon}},
\code{\link{termemot}},
\code{\link{terminal}},
\code{\link{termpain}},
\code{\link{termpay}},
\code{\link{termrel}}
}
\concept{Mental Health}
\keyword{variable}
