% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_249.R
\name{socsecnu}
\alias{socsecnu}
\title{Social security phase out or continue}
\source{
General Social Survey https://gss.norc.org
}
\description{
socsecnu
}
\details{
Question Do you think it would be better to phase out Social Security and require people to take what they and their employers now pay into Social Security and invest it in mandatory private retirement accounts, or do you think Social Security should continue much as it is now?
}
\section{Values}{

\itemize{
\item \code{1} phase out social security+
\item \code{2} continue as is
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2244/vshow}{GSS Data Explorer page for \code{socsecnu}}.
}

\seealso{
Other Mental Health: 
\code{\link{prozkid3}},
\code{\link{socsecfx}},
\code{\link{socsecrt}},
\code{\link{termecon}},
\code{\link{termemot}},
\code{\link{terminal}},
\code{\link{termpain}},
\code{\link{termpay}},
\code{\link{termrel}}

Other Federal Government: 
\code{\link{socsecfx}},
\code{\link{socsecrt}}

Other Social Security: 
\code{\link{socsecfx}},
\code{\link{socsecrt}}

Other United States: 
\code{\link{socsecfx}},
\code{\link{socsecrt}}
}
\concept{Federal Government}
\concept{Mental Health}
\concept{Social Security}
\concept{United States}
\keyword{variable}
