% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_59.R
\name{socsibs}
\alias{socsibs}
\title{Spend evening with siblings}
\source{
General Social Survey https://gss.norc.org
}
\description{
socsibs
}
\details{
Question Would you use this card and tell me which answer comes closest to how often you do the following things…
F. Spend a social evening with a brother or sister?
}
\section{Values}{

\itemize{
\item \code{0} no such people
\item \code{1} almost daily
\item \code{2} sev times a week
\item \code{3} sev times a mnth
\item \code{4} once a month
\item \code{5} sev times a year
\item \code{6} once a year
\item \code{7} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1978 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1988 \tab A/B \tab partial \cr
   1989 \tab A/B \tab partial \cr
   1990 \tab A/B \tab partial \cr
   1991 \tab A/B \tab partial \cr
   1993 \tab A/B \tab partial \cr
   1994 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/526/vshow}{GSS Data Explorer page for \code{socsibs}}.
}

\seealso{
Other Core: 
\code{\link{aged}},
\code{\link{agedchld}},
\code{\link{agedpar}},
\code{\link{drink}},
\code{\link{drunk}},
\code{\link{partfull}},
\code{\link{smoke}},
\code{\link{socpars}},
\code{\link{weekswrk}}

Other Family: 
\code{\link{socpars}}

Other Sociability: 
\code{\link{socpars}}
}
\concept{Core}
\concept{Family}
\concept{Sociability}
\keyword{variable}
