% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_476.R
\name{soctype1}
\alias{soctype1}
\title{What type of society is america}
\source{
General Social Survey https://gss.norc.org
}
\description{
soctype1
}
\details{
Question These five diagrams show different types of society. Please read the descriptions and look at the diagrams and decide which you think best describes America today. . .
}
\section{Values}{

\itemize{
\item \verb{[1]} type a
\item \verb{[2]} type b
\item \verb{[3]} type c
\item \verb{[4]} type d
\item \verb{[5]} type e
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   B/C/- \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4336/vshow}{GSS Data Explorer page for \code{soctype1}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab type a \tab type b \tab type c \tab type d \tab type e \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2000 \tab 1545 \tab 193 \tab 364 \tab 213 \tab 322 \tab 32 \tab 64 \tab 84 \tab - \tab 2817 \cr
   2021 \tab 2029 \tab 436 \tab 680 \tab 268 \tab 307 \tab 48 \tab 73 \tab 177 \tab 14 \tab 4032 \cr
   Total \tab 3574 \tab 629 \tab 1044 \tab 481 \tab 629 \tab 80 \tab 137 \tab 261 \tab 14 \tab 6849 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{execrnk}},
\code{\link{famrnk}},
\code{\link{justpay}},
\code{\link{pasup}},
\code{\link{paychild}},
\code{\link{paydowel}},
\code{\link{payhard}},
\code{\link{soctype2}},
\code{\link{unsklrnk}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
