% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_476.R
\name{soctype1}
\alias{soctype1}
\title{What type of society is america}
\source{
General Social Survey https://gss.norc.org
}
\description{
soctype1
}
\details{
Question These five diagrams show different types of society. Please read the descriptions and look at the diagrams and decide which you think best describes America today. . .
}
\section{Values}{

\itemize{
\item \code{1} type a
\item \code{2} type b
\item \code{3} type c
\item \code{4} type d
\item \code{5} type e
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2021 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4336/vshow}{GSS Data Explorer page for \code{soctype1}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{execrnk}},
\code{\link{famrnk}},
\code{\link{justpay}},
\code{\link{pasup}},
\code{\link{paychild}},
\code{\link{paydowel}},
\code{\link{payhard}},
\code{\link{soctype2}},
\code{\link{unsklrnk}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
