% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_476.R
\name{soctype2}
\alias{soctype2}
\title{What type of society does r prefer}
\source{
General Social Survey https://gss.norc.org
}
\description{
soctype2
}
\details{
Question What do you think America ought to be like - which would you prefer?
}
\section{Values}{

\itemize{
\item \verb{[1]} type a
\item \verb{[2]} type b
\item \verb{[3]} type c
\item \verb{[4]} type d
\item \verb{[5]} type e
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   B/C/- \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4337/vshow}{GSS Data Explorer page for \code{soctype2}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab type a \tab type b \tab type c \tab type d \tab type e \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2000 \tab 1545 \tab 26 \tab 101 \tab 177 \tab 514 \tab 280 \tab 88 \tab 86 \tab - \tab 2817 \cr
   2021 \tab 2029 \tab 27 \tab 87 \tab 194 \tab 965 \tab 417 \tab 107 \tab 178 \tab 28 \tab 4032 \cr
   Total \tab 3574 \tab 53 \tab 188 \tab 371 \tab 1479 \tab 697 \tab 195 \tab 264 \tab 28 \tab 6849 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{execrnk}},
\code{\link{famrnk}},
\code{\link{justpay}},
\code{\link{pasup}},
\code{\link{paychild}},
\code{\link{paydowel}},
\code{\link{payhard}},
\code{\link{soctype1}},
\code{\link{unsklrnk}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
