% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_108.R
\name{solfarm}
\alias{solfarm}
\title{Farm org tries to solve problems}
\source{
General Social Survey https://gss.norc.org
}
\description{
solfarm
}
\details{
Question Do (EACH "YES" IN  MEMFRAT-MEMNUM) to which you belong do anything to try to solve individual or community problems?
L. Farm organization
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/991/vshow}{GSS Data Explorer page for \code{solfarm}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1987 \tab 1743 \tab 43 \tab 7 \tab 1 \tab 25 \tab 1819 \cr
   Total \tab 1743 \tab 43 \tab 7 \tab 1 \tab 25 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{actfrat}},
\code{\link{actpolit}},
\code{\link{actserv}},
\code{\link{actvet}},
\code{\link{solchurh}},
\code{\link{sollit}},
\code{\link{solnat}},
\code{\link{solother}},
\code{\link{solprof}}
}
\concept{Socio-political participation}
\keyword{variable}
