% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_540.R
\name{solok}
\alias{solok}
\title{How important: citizens have adequate standard of living}
\source{
General Social Survey https://gss.norc.org
}
\description{
solok
}
\details{
Question There are different opinions about people's rights in a democracy. On a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it:
A. That all citizens have an adequate standard of living
}
\section{Values}{

\itemize{
\item \code{1} not at all important
\item \code{7} very important
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4933/vshow}{GSS Data Explorer page for \code{solok}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab not at all important \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 10 \tab 9 \tab 22 \tab 83 \tab 184 \tab 200 \tab 954 \tab 5 \tab 5 \tab 2812 \cr
   2014 \tab 1274 \tab 13 \tab 7 \tab 24 \tab 95 \tab 179 \tab 201 \tab 714 \tab 29 \tab 2 \tab 2538 \cr
   Total \tab 2614 \tab 23 \tab 16 \tab 46 \tab 178 \tab 363 \tab 401 \tab 1668 \tab 34 \tab 7 \tab 5350 \cr
}
}

\seealso{
Other ISSP Citizenship: 
\code{\link{citviews}},
\code{\link{eqtreat}},
\code{\link{grpother}},
\code{\link{grpparty}},
\code{\link{grprelig}},
\code{\link{grpsprts}},
\code{\link{grpwork}},
\code{\link{polopts}},
\code{\link{rghtsmin}}
}
\concept{ISSP Citizenship}
\keyword{variable}
