% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_333.R
\name{solveown}
\alias{solveown}
\title{R would solve problem on ones own or get advice}
\source{
General Social Survey https://gss.norc.org
}
\description{
solveown
}
\details{
Question When you have a problem, do you prefer to solve it on your own or do you like to talk to other people to get advice, help or information?
}
\section{Values}{

\itemize{
\item \code{1} on own
\item \code{2} get advice
\item \code{3} both(on own and get advice) depending on problem
\item \code{4} religious
\item \code{5} research the problem
\item \code{6} talk to a doctor other professional
\item \code{7} talk with older people
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2978/vshow}{GSS Data Explorer page for \code{solveown}}.
}

\seealso{
Other Mental Health II: 
\code{\link{doc15a}},
\code{\link{doc16a}},
\code{\link{doc18a}},
\code{\link{doc19a}},
\code{\link{doc20a}},
\code{\link{medaddct}},
\code{\link{medsymps}},
\code{\link{medunacc}},
\code{\link{medweak}}
}
\concept{Life Events}
\concept{Mental Health II}
\keyword{variable}
