% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_333.R
\name{solveown}
\alias{solveown}
\title{R would solve problem on ones own or get advice}
\source{
General Social Survey https://gss.norc.org
}
\description{
solveown
}
\details{
Question When you have a problem, do you prefer to solve it on your own or do you like to talk to other people to get advice, help or information?
}
\section{Values}{

\itemize{
\item \verb{[1]} on own
\item \verb{[2]} get advice
\item \verb{[3]} both(on own and get advice) depending on problem
\item \verb{[4]} religious
\item \verb{[5]} research the problem
\item \verb{[6]} talk to a doctor other professional
\item \verb{[7]} talk with older people
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2978/vshow}{GSS Data Explorer page for \code{solveown}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab on own \tab get advice \tab both(on own and get advice) depending on problem \tab religious \tab research the problem \tab talk to a doctor other professional \tab talk with older people \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 593 \tab 729 \tab 35 \tab 7 \tab 3 \tab 4 \tab 1 \tab 14 \tab 7 \tab 2765 \cr
   Total \tab 1372 \tab 593 \tab 729 \tab 35 \tab 7 \tab 3 \tab 4 \tab 1 \tab 14 \tab 7 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{doc15a}},
\code{\link{doc16a}},
\code{\link{doc18a}},
\code{\link{doc19a}},
\code{\link{doc20a}},
\code{\link{medaddct}},
\code{\link{medsymps}},
\code{\link{medunacc}},
\code{\link{medweak}}
}
\concept{Mental Health II}
\keyword{variable}
