% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_452.R
\name{sonnum}
\alias{sonnum}
\title{Number of adult sons}
\source{
General Social Survey https://gss.norc.org
}
\description{
sonnum
}
\details{
Question How many sons age 18 or older do you have? (We mean sons who are still alive; please include step-sons, half-sons, and adopted sons.)
}
\section{Values}{

\itemize{
\item \verb{[0]} none
\item \verb{[1]} one
\item \verb{[2]} two
\item \verb{[3]} three
\item \verb{[4]} four
\item \verb{[5]} five +
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4147/vshow}{GSS Data Explorer page for \code{sonnum}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab none \tab one \tab two \tab three \tab four \tab five + \tab no answer \tab Total \cr
   1986 \tab - \tab 973 \tab 237 \tab 150 \tab 55 \tab 30 \tab 22 \tab 3 \tab 1470 \cr
   Total \tab 0 \tab 973 \tab 237 \tab 150 \tab 55 \tab 30 \tab 22 \tab 3 \tab 1470 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{brocall}},
\code{\link{daucall}},
\code{\link{daunum}},
\code{\link{dautime}},
\code{\link{dauvisit}},
\code{\link{posslq}},
\code{\link{soncall}},
\code{\link{sontime}},
\code{\link{sonvisit}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
