% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_452.R
\name{sontime}
\alias{sontime}
\title{Travel time to son}
\source{
General Social Survey https://gss.norc.org
}
\description{
sontime
}
\details{
Question How many sons age 18 or older do you have? (We mean sons who are still alive; please include step-sons, half-sons, and adopted sons.)
B. About how long would it take you to get to where your son lives? Think of the time it usually takes door to door.
}
\section{Values}{

\itemize{
\item \code{1} lt 15 min
\item \code{2} 15-30 min
\item \code{3} 30 min-1 hr
\item \code{4} 1-2 hr
\item \code{5} 2-3 hr
\item \code{6} 3-5 hr
\item \code{7} 5-12 hr
\item \code{8} 12+ hr
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1986 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4149/vshow}{GSS Data Explorer page for \code{sontime}}.
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{brocall}},
\code{\link{daucall}},
\code{\link{daunum}},
\code{\link{dautime}},
\code{\link{dauvisit}},
\code{\link{posslq}},
\code{\link{soncall}},
\code{\link{sonnum}},
\code{\link{sonvisit}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
