% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_442.R
\name{sparms}
\alias{sparms}
\title{Govt spending on defense}
\source{
General Social Survey https://gss.norc.org
}
\description{
sparms
}
\details{
Question Listed below are various areas of government spending. Please indicate whether you would like to see more or less government spending in each area. Remember that if you say "much more," it might require a tax increase to pay for it.
E. The military and defense.
}
\section{Values}{

\itemize{
\item \verb{[1]} spend much more
\item \verb{[2]} spend more
\item \verb{[3]} spend same
\item \verb{[4]} spend less
\item \verb{[5]} spend much less
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1990, 1996 \cr
   D/-/- \tab 2006 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4061/vshow}{GSS Data Explorer page for \code{sparms}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab spend much more \tab spend more \tab spend same \tab spend less \tab spend much less \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 49 \tab 77 \tab 237 \tab 183 \tab 92 \tab 21 \tab 18 \tab 1534 \cr
   1990 \tab 155 \tab 37 \tab 120 \tab 422 \tab 340 \tab 217 \tab 36 \tab 45 \tab 1372 \cr
   1996 \tab 1572 \tab 61 \tab 211 \tab 585 \tab 315 \tab 106 \tab 19 \tab 35 \tab 2904 \cr
   2006 \tab 2992 \tab 189 \tab 356 \tab 542 \tab 315 \tab 95 \tab 18 \tab 3 \tab 4510 \cr
   2016 \tab 1477 \tab 184 \tab 421 \tab 479 \tab 196 \tab 87 \tab 12 \tab 11 \tab 2867 \cr
   Total \tab 7053 \tab 520 \tab 1185 \tab 2265 \tab 1349 \tab 597 \tab 106 \tab 112 \tab 13187 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{bustax}},
\code{\link{inctax}},
\code{\link{infljobs}},
\code{\link{sparts}},
\code{\link{sphlth}},
\code{\link{sppolice}},
\code{\link{spretire}},
\code{\link{spschool}},
\code{\link{spunemp}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
