% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_442.R
\name{sparts}
\alias{sparts}
\title{Govt spending on culture and the arts}
\source{
General Social Survey https://gss.norc.org
}
\description{
sparts
}
\details{
Question Listed below are various areas of government spending. Please indicate whether you would like to see more or less government spending in each area. Remember that if you say "much more," it might require a tax increase to pay for it.
H. Culture and the arts.
}
\section{Values}{

\itemize{
\item \verb{[1]} spend much more
\item \verb{[2]} spend more
\item \verb{[3]} spend same
\item \verb{[4]} spend less
\item \verb{[5]} spend much less
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1990, 1996 \cr
   D/-/- \tab 2006 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4064/vshow}{GSS Data Explorer page for \code{sparts}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab spend much more \tab spend more \tab spend same \tab spend less \tab spend much less \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 27 \tab 67 \tab 257 \tab 162 \tab 101 \tab 48 \tab 15 \tab 1534 \cr
   1990 \tab 155 \tab 27 \tab 122 \tab 488 \tab 246 \tab 191 \tab 99 \tab 44 \tab 1372 \cr
   1996 \tab 1572 \tab 55 \tab 140 \tab 473 \tab 345 \tab 225 \tab 55 \tab 39 \tab 2904 \cr
   2006 \tab 2992 \tab 72 \tab 288 \tab 686 \tab 337 \tab 105 \tab 30 \tab - \tab 4510 \cr
   2016 \tab 1477 \tab 39 \tab 254 \tab 653 \tab 328 \tab 84 \tab 21 \tab 11 \tab 2867 \cr
   Total \tab 7053 \tab 220 \tab 871 \tab 2557 \tab 1418 \tab 706 \tab 253 \tab 109 \tab 13187 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{bustax}},
\code{\link{inctax}},
\code{\link{infljobs}},
\code{\link{sparms}},
\code{\link{sphlth}},
\code{\link{sppolice}},
\code{\link{spretire}},
\code{\link{spschool}},
\code{\link{spunemp}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
