% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_35.R
\name{spattend}
\alias{spattend}
\title{How often sp attends religious services}
\source{
General Social Survey https://gss.norc.org
}
\description{
spattend
}
\details{
Question C. How often does your (husband/wife) attend religious services?
(USE CATEGORIES AS PROBES, IF NECESSARY.)
}
\section{Values}{

\itemize{
\item \code{0} never
\item \code{1} lt once a year
\item \code{2} once a year
\item \code{3} sevrl times a yr
\item \code{4} once a month
\item \code{5} 2-3x a month
\item \code{6} nrly every week
\item \code{7} every week
\item \code{8} more thn once wk
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1989 \cr
   A/B/C \tab 1988 \cr
   Full \tab 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/297/vshow}{GSS Data Explorer page for \code{spattend}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab never \tab lt once a year \tab once a year \tab sevrl times a yr \tab once a month \tab 2-3x a month \tab nrly every week \tab every week \tab more thn once wk \tab don't know \tab Total \cr
   1986 \tab 644 \tab 147 \tab 53 \tab 80 \tab 90 \tab 54 \tab 67 \tab 46 \tab 209 \tab 77 \tab 3 \tab 1470 \cr
   1987 \tab 903 \tab 151 \tab 51 \tab 108 \tab 121 \tab 71 \tab 81 \tab 52 \tab 200 \tab 71 \tab 10 \tab 1819 \cr
   1988 \tab 693 \tab 148 \tab 47 \tab 81 \tab 91 \tab 50 \tab 57 \tab 57 \tab 187 \tab 66 \tab 4 \tab 1481 \cr
   1989 \tab 976 \tab 110 \tab 31 \tab 60 \tab 64 \tab 32 \tab 39 \tab 31 \tab 148 \tab 42 \tab 4 \tab 1537 \cr
   Total \tab 3216 \tab 556 \tab 182 \tab 329 \tab 366 \tab 207 \tab 244 \tab 186 \tab 744 \tab 256 \tab 21 \tab 6307 \cr
}
}

\seealso{
Other Core: 
\code{\link{attend}},
\code{\link{fund}},
\code{\link{jew}},
\code{\link{jewaj}},
\code{\link{maattend}},
\code{\link{paattend}},
\code{\link{postlife}},
\code{\link{reliten}},
\code{\link{relitena}}
}
\concept{Core}
\keyword{variable}
