% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_486.R
\name{spborn}
\alias{spborn}
\title{When was your spouse born}
\source{
General Social Survey https://gss.norc.org
}
\description{
spborn
}
\details{
Question When was your spouse/steady partner born?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4440/vshow}{GSS Data Explorer page for \code{spborn}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1918 \tab 1923 \tab 1925 \tab 1926 \tab 1927 \tab 1929 \tab 1930 \tab 1931 \tab 1932 \tab 1933 \tab 1934 \tab 1935 \tab 1936 \tab 1937 \tab 1938 \tab 1939 \tab 1940 \tab 1941 \tab 1942 \tab 1943 \tab 1944 \tab 1945 \tab 1946 \tab 1947 \tab 1948 \tab 1949 \tab 1950 \tab 1951 \tab 1952 \tab 1953 \tab 1954 \tab 1955 \tab 1956 \tab 1957 \tab 1958 \tab 1959 \tab 1960 \tab 1961 \tab 1962 \tab 1963 \tab 1964 \tab 1965 \tab 1966 \tab 1967 \tab 1968 \tab 1969 \tab 1970 \tab 1971 \tab 1972 \tab 1973 \tab 1974 \tab 1975 \tab 1976 \tab 1977 \tab 1978 \tab 1979 \tab 1980 \tab 1981 \tab 1982 \tab 1983 \tab 1984 \tab 1985 \tab 1986 \tab 1987 \tab 1988 \tab 1989 \tab 1990 \tab 1991 \tab 1992 \tab 9998 \tab 9999 \tab Total \cr
   2012 \tab 1274 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 2 \tab 7 \tab 5 \tab 4 \tab 4 \tab 3 \tab 3 \tab 7 \tab 5 \tab 2 \tab 1 \tab 10 \tab 13 \tab 5 \tab 8 \tab 4 \tab 6 \tab 11 \tab 13 \tab 5 \tab 10 \tab 14 \tab 10 \tab 9 \tab 10 \tab 11 \tab 19 \tab 12 \tab 13 \tab 15 \tab 18 \tab 9 \tab 19 \tab 16 \tab 15 \tab 13 \tab 13 \tab 9 \tab 17 \tab 13 \tab 10 \tab 12 \tab 14 \tab 19 \tab 17 \tab 14 \tab 14 \tab 25 \tab 11 \tab 17 \tab 15 \tab 16 \tab 19 \tab 12 \tab 3 \tab 18 \tab 9 \tab 8 \tab 8 \tab 6 \tab 7 \tab 5 \tab 1 \tab 9 \tab 18 \tab 1974 \cr
   Total \tab 1274 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 2 \tab 2 \tab 7 \tab 5 \tab 4 \tab 4 \tab 3 \tab 3 \tab 7 \tab 5 \tab 2 \tab 1 \tab 10 \tab 13 \tab 5 \tab 8 \tab 4 \tab 6 \tab 11 \tab 13 \tab 5 \tab 10 \tab 14 \tab 10 \tab 9 \tab 10 \tab 11 \tab 19 \tab 12 \tab 13 \tab 15 \tab 18 \tab 9 \tab 19 \tab 16 \tab 15 \tab 13 \tab 13 \tab 9 \tab 17 \tab 13 \tab 10 \tab 12 \tab 14 \tab 19 \tab 17 \tab 14 \tab 14 \tab 25 \tab 11 \tab 17 \tab 15 \tab 16 \tab 19 \tab 12 \tab 3 \tab 18 \tab 9 \tab 8 \tab 8 \tab 6 \tab 7 \tab 5 \tab 1 \tab 9 \tab 18 \tab 1974 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{caresik1}},
\code{\link{clean1}},
\code{\link{cooking1}},
\code{\link{deckids}},
\code{\link{hhwkdis}},
\code{\link{hhwkfair}},
\code{\link{rhhwork}},
\code{\link{shop1}},
\code{\link{sphhwork}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
