% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_577.R
\name{spdays}
\alias{spdays}
\title{Days spouse normally works}
\source{
General Social Survey https://gss.norc.org
}
\description{
spdays
}
\details{
Question What days of the week does (SPOUSE) normally work?  CODE ALL THAT APPLY.
}
\section{Values}{

\itemize{
\item \code{1} mon to fri
\item \code{2} mon to fri,sat
\item \code{3} mon to fri,sun
\item \code{4} wkend plus
\item \code{5} no schedule
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5289/vshow}{GSS Data Explorer page for \code{spdays}}.
}

\seealso{
Other Military Service & Work: 
\code{\link{sphour}},
\code{\link{vetkind}},
\code{\link{vetyears}},
\code{\link{workdy}},
\code{\link{workhr}}
}
\concept{Military Service & Work}
\keyword{variable}
