% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_41.R
\name{spden16}
\alias{spden16}
\title{Denomination in which spouse raised}
\source{
General Social Survey https://gss.norc.org
}
\description{
spden16
}
\details{
Question In what religion was your (husband/wife) raised?
A. What specific denomination is that, if any? (PROBE FOR EXACT DENOMINATION)
}
\section{Values}{

\itemize{
\item \code{10} am baptist asso
\item \code{11} am bapt ch in usa
\item \code{12} nat bapt conv of am
\item \code{13} nat bapt conv usa
\item \code{14} southern baptist
\item \code{15} other baptists
\item \code{18} baptist-dk which
\item \code{20} afr meth episcopal
\item \code{21} afr meth ep zion
\item \code{22} united methodist
\item \code{23} other methodist
\item \code{28} methodist-dk which
\item \code{30} am lutheran
\item \code{31} luth ch in america
\item \code{32} lutheran-mo synod
\item \code{33} wi evan luth synod
\item \code{34} other lutheran
\item \code{35} evangelic luth
\item \code{38} lutheran-dk which
\item \code{40} presbyterian c in us
\item \code{41} united pres ch in usa
\item \code{42} other presbyterian
\item \code{43} presbyterian, merged
\item \code{48} presbyterian-dk wh
\item \code{50} episcopal
\item \code{60} other
\item \code{70} no denomination
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/-/- \tab full \cr
   1974 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1976 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/356/vshow}{GSS Data Explorer page for \code{spden16}}.
}

\seealso{
Other Core: 
\code{\link{corel}},
\code{\link{spfund}},
\code{\link{spfund16}},
\code{\link{spjew}},
\code{\link{spjew16}},
\code{\link{spjewaj}},
\code{\link{spoth16}},
\code{\link{spothjew}},
\code{\link{sprel16}}
}
\concept{Core}
\keyword{variable}
