% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_12.R
\name{speftotlt}
\alias{speftotlt}
\title{Total number of students enrolled in college spouse attended}
\source{
General Social Survey https://gss.norc.org
}
\description{
speftotlt
}
\details{
Question In what year did you receive the (COLLEGE NAME) degree?
J. Total Fall Enrollment
}
\section{Values}{

\itemize{
\item \verb{[1]} 0-467 students enrolled
\item \verb{[2]} 468 or more students
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/78/vshow}{GSS Data Explorer page for \code{speftotlt}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab 0-467 students enrolled \tab 468 or more students \tab no answer \tab Total \cr
   2012 \tab 1706 \tab 2 \tab 265 \tab 1 \tab 1974 \cr
   2014 \tab 2123 \tab 5 \tab 410 \tab - \tab 2538 \cr
   2016 \tab 2460 \tab 5 \tab 402 \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab 2348 \cr
   Total \tab 8637 \tab 12 \tab 1077 \tab 1 \tab 9727 \cr
}
}

\seealso{
Other Core: 
\code{\link{barate}},
\code{\link{eftotlt}},
\code{\link{gradtounder}},
\code{\link{sector}},
\code{\link{spsector}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme1}},
\code{\link{voednme2}}
}
\concept{Core}
\keyword{variable}
