% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_574.R
\name{spethnic}
\alias{spethnic}
\title{Country of spouse's family origin}
\source{
General Social Survey https://gss.norc.org
}
\description{
spethnic
}
\details{
Question From what countries or part of the world did your (wife''s/husband''s) ancestors come?
A. IF MORE THAN ONE COUNTRY NAMED:  Which one of these countries does your (wife/husband) feel closer to?
A. SUMMARY ETHNICITY OF SPOUSE
}
\section{Values}{

\itemize{
\item \code{1} africa
\item \code{2} austria
\item \code{3} french canada
\item \code{4} other canada
\item \code{5} china
\item \code{6} czechoslovakia
\item \code{7} denmark
\item \code{8} england & wales
\item \code{9} finland
\item \code{10} france
\item \code{11} germany
\item \code{12} greece
\item \code{13} hungary
\item \code{14} ireland
\item \code{15} italy
\item \code{16} japan
\item \code{17} mexico
\item \code{18} netherlands
\item \code{19} norway
\item \code{20} philippines
\item \code{21} poland
\item \code{22} puerto rico
\item \code{23} russia
\item \code{24} scotland
\item \code{25} spain
\item \code{26} sweden
\item \code{27} switzerland
\item \code{28} west indies
\item \code{29} other
\item \code{30} american indian
\item \code{31} india
\item \code{32} portugal
\item \code{33} lithuania
\item \code{34} yugoslavia
\item \code{35} rumania
\item \code{36} belgium
\item \code{37} arabic
\item \code{38} other spanish
\item \code{39} non-span windies
\item \code{40} other asian
\item \code{41} other european
\item \code{97} american only
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
   Full \tab 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5268/vshow}{GSS Data Explorer page for \code{spethnic}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab uncodeable \tab africa \tab austria \tab french canada \tab other canada \tab china \tab czechoslovakia \tab denmark \tab england & wales \tab finland \tab france \tab germany \tab greece \tab ireland \tab italy \tab japan \tab mexico \tab netherlands \tab norway \tab philippines \tab poland \tab puerto rico \tab russia \tab scotland \tab spain \tab sweden \tab switzerland \tab other \tab american indian \tab india \tab portugal \tab lithuania \tab yugoslavia \tab rumania \tab arabic \tab other spanish \tab non-span windies \tab american only \tab no answer \tab hungary \tab west indies \tab belgium \tab other asian \tab other european \tab not available in this year \tab iap \tab Total \cr
   1985 \tab 853 \tab 25 \tab 10 \tab 12 \tab 5 \tab 2 \tab 10 \tab 7 \tab 104 \tab 5 \tab 21 \tab 120 \tab 4 \tab 87 \tab 35 \tab 4 \tab 27 \tab 13 \tab 18 \tab 2 \tab 27 \tab 5 \tab 13 \tab 26 \tab 2 \tab 16 \tab 6 \tab 5 \tab 36 \tab 1 \tab 3 \tab 2 \tab 2 \tab 3 \tab 1 \tab 4 \tab 2 \tab 11 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 818 \tab 34 \tab 7 \tab 9 \tab 6 \tab - \tab 7 \tab 9 \tab 100 \tab 3 \tab 11 \tab 121 \tab 2 \tab 86 \tab 41 \tab 3 \tab 23 \tab 9 \tab 15 \tab 1 \tab 22 \tab 8 \tab 10 \tab 21 \tab 5 \tab 12 \tab 6 \tab 6 \tab 22 \tab 5 \tab 3 \tab 3 \tab 5 \tab 1 \tab 1 \tab 3 \tab 3 \tab 15 \tab 4 \tab 7 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1134 \tab 84 \tab 7 \tab 13 \tab 2 \tab 4 \tab 14 \tab 10 \tab 78 \tab - \tab 17 \tab 111 \tab 1 \tab 63 \tab 47 \tab 2 \tab 23 \tab 13 \tab 15 \tab 1 \tab 30 \tab 1 \tab 12 \tab 24 \tab 3 \tab 15 \tab 3 \tab 11 \tab 34 \tab 2 \tab 3 \tab 2 \tab 2 \tab 2 \tab 2 \tab 1 \tab 3 \tab 14 \tab 10 \tab 5 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 880 \tab 29 \tab 3 \tab 10 \tab 3 \tab 7 \tab 7 \tab 4 \tab 85 \tab 5 \tab 12 \tab 116 \tab 7 \tab 67 \tab 35 \tab 2 \tab 22 \tab 25 \tab 10 \tab - \tab 17 \tab 5 \tab 15 \tab 18 \tab 9 \tab 10 \tab 5 \tab 8 \tab 28 \tab 3 \tab - \tab - \tab 2 \tab 1 \tab - \tab 2 \tab 1 \tab 13 \tab 10 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 888 \tab 32 \tab 2 \tab 12 \tab 4 \tab 7 \tab 8 \tab 11 \tab 87 \tab - \tab 13 \tab 139 \tab 4 \tab 75 \tab 43 \tab 4 \tab 25 \tab 11 \tab 15 \tab 3 \tab 28 \tab 5 \tab 6 \tab 24 \tab 5 \tab 9 \tab 2 \tab 12 \tab 26 \tab 2 \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 6 \tab 15 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 802 \tab 34 \tab 6 \tab 7 \tab 4 \tab 1 \tab 12 \tab 7 \tab 80 \tab 3 \tab 13 \tab 99 \tab 2 \tab 64 \tab 32 \tab 3 \tab 23 \tab 14 \tab 17 \tab - \tab 21 \tab 7 \tab 8 \tab 21 \tab 7 \tab 7 \tab 5 \tab 6 \tab 26 \tab 1 \tab 3 \tab - \tab 4 \tab 1 \tab 1 \tab 4 \tab 6 \tab 6 \tab 13 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 879 \tab 46 \tab 2 \tab 15 \tab 9 \tab 3 \tab 7 \tab 12 \tab 88 \tab 1 \tab 17 \tab 109 \tab 2 \tab 78 \tab 46 \tab 4 \tab 25 \tab 10 \tab 8 \tab 3 \tab 22 \tab 5 \tab 10 \tab 15 \tab 5 \tab 14 \tab 4 \tab 13 \tab 24 \tab 3 \tab 1 \tab 3 \tab 2 \tab - \tab - \tab 2 \tab 6 \tab 6 \tab 12 \tab 5 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 939 \tab 37 \tab 1 \tab 6 \tab 4 \tab 3 \tab 6 \tab 6 \tab 103 \tab 2 \tab 16 \tab 143 \tab 8 \tab 64 \tab 33 \tab 2 \tab 13 \tab 13 \tab 9 \tab 8 \tab 23 \tab 1 \tab 18 \tab 21 \tab 15 \tab 16 \tab 2 \tab 14 \tab 28 \tab 3 \tab 2 \tab 2 \tab 1 \tab 2 \tab 1 \tab 9 \tab 3 \tab 9 \tab 16 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2659 \tab 26 \tab 2 \tab 10 \tab 2 \tab - \tab 5 \tab 1 \tab 38 \tab 2 \tab 6 \tab 71 \tab 2 \tab 33 \tab 15 \tab 3 \tab 19 \tab 6 \tab 7 \tab 2 \tab 7 \tab 2 \tab 4 \tab 13 \tab - \tab 8 \tab 2 \tab 3 \tab 7 \tab 3 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab - \tab 5 \tab 13 \tab 2 \tab - \tab - \tab 4 \tab 1 \tab - \tab - \tab 2992 \cr
   Total \tab 9852 \tab 347 \tab 40 \tab 94 \tab 39 \tab 27 \tab 76 \tab 67 \tab 763 \tab 21 \tab 126 \tab 1029 \tab 32 \tab 617 \tab 327 \tab 27 \tab 200 \tab 114 \tab 114 \tab 20 \tab 197 \tab 39 \tab 96 \tab 183 \tab 51 \tab 107 \tab 35 \tab 78 \tab 231 \tab 23 \tab 19 \tab 17 \tab 20 \tab 12 \tab 8 \tab 31 \tab 25 \tab 85 \tab 98 \tab 30 \tab 1 \tab 5 \tab 4 \tab 1 \tab 0 \tab 0 \tab 15328 \cr
}
}

\seealso{
Other Ethnicity: 
\code{\link{eth1}},
\code{\link{eth2}},
\code{\link{eth3}},
\code{\link{ethnic}},
\code{\link{ethnum}},
\code{\link{speth1}},
\code{\link{speth2}},
\code{\link{speth3}},
\code{\link{spethnum}}
}
\concept{Ethnicity}
\keyword{variable}
