% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_574.R
\name{spethnum}
\alias{spethnum}
\title{Type of response about ethnicity:sp}
\source{
General Social Survey https://gss.norc.org
}
\description{
spethnum
}
\details{
Question (RECODE OF SPETHNIC)
}
\section{Values}{

\itemize{
\item \code{1} names 1
\item \code{2} chooses 1 of 2+
\item \code{3} cannot choose 1
\item \code{4} cannot name any
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5272/vshow}{GSS Data Explorer page for \code{spethnum}}.
}

\seealso{
Other Ethnicity: 
\code{\link{eth1}},
\code{\link{eth2}},
\code{\link{eth3}},
\code{\link{ethnic}},
\code{\link{ethnum}},
\code{\link{speth1}},
\code{\link{speth2}},
\code{\link{speth3}},
\code{\link{spethnic}}
}
\concept{Ethnicity}
\keyword{variable}
