% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_41.R
\name{spfund16}
\alias{spfund16}
\title{How fundamentalist was spouse at 16}
\source{
General Social Survey https://gss.norc.org
}
\description{
spfund16
}
\details{
Question In what religion was your (husband/wife) raised?
C. Fundamentalism/Liberalism of Religion Spouse Raised In
}
\section{Values}{

\itemize{
\item \code{1} fundamentalist
\item \code{2} moderate
\item \code{3} liberal
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/-/- \tab full \cr
   1974 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1976 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/C \tab full \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/359/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab fundamentalist \tab liberal \tab moderate \tab no answer \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 428 \tab 321 \tab 145 \tab 577 \tab 33 \tab - \tab 1504 \cr
   1974 \tab 419 \tab 301 \tab 160 \tab 579 \tab 25 \tab - \tab 1484 \cr
   1975 \tab 488 \tab 299 \tab 142 \tab 541 \tab 20 \tab - \tab 1490 \cr
   1976 \tab 525 \tab 279 \tab 133 \tab 539 \tab 23 \tab - \tab 1499 \cr
   1977 \tab 555 \tab 302 \tab 138 \tab 509 \tab 26 \tab - \tab 1530 \cr
   1978 \tab 572 \tab 284 \tab 143 \tab 504 \tab 29 \tab - \tab 1532 \cr
   1980 \tab 579 \tab 270 \tab 129 \tab 435 \tab 55 \tab - \tab 1468 \cr
   1982 \tab 867 \tab 339 \tab 133 \tab 495 \tab 26 \tab - \tab 1860 \cr
   1983 \tab 634 \tab 275 \tab 157 \tab 506 \tab 27 \tab - \tab 1599 \cr
   1984 \tab 644 \tab 275 \tab 175 \tab 348 \tab 31 \tab - \tab 1473 \cr
   1985 \tab 663 \tab 293 \tab 205 \tab 350 \tab 23 \tab - \tab 1534 \cr
   1986 \tab 644 \tab 270 \tab 193 \tab 319 \tab 44 \tab - \tab 1470 \cr
   1987 \tab 903 \tab 327 \tab 221 \tab 336 \tab 32 \tab - \tab 1819 \cr
   1988 \tab 693 \tab 263 \tab 205 \tab 289 \tab 31 \tab - \tab 1481 \cr
   1989 \tab 690 \tab 265 \tab 179 \tab 367 \tab 36 \tab - \tab 1537 \cr
   1990 \tab 645 \tab 195 \tab 180 \tab 324 \tab 28 \tab - \tab 1372 \cr
   1991 \tab 713 \tab 256 \tab 179 \tab 346 \tab 23 \tab - \tab 1517 \cr
   1993 \tab 747 \tab 269 \tab 202 \tab 345 \tab 43 \tab - \tab 1606 \cr
   1994 \tab 2213 \tab 217 \tab 174 \tab 308 \tab 80 \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 51701 \tab 5300 \tab 3193 \tab 8017 \tab 635 \tab 6853 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{corel}},
\code{\link{spden16}},
\code{\link{spfund}},
\code{\link{spjew}},
\code{\link{spjew16}},
\code{\link{spjewaj}},
\code{\link{spoth16}},
\code{\link{spothjew}},
\code{\link{sprel16}}

Other Religion: 
\code{\link{spden16}},
\code{\link{spfund}},
\code{\link{spjew}},
\code{\link{spjew16}},
\code{\link{spoth16}},
\code{\link{spothjew}},
\code{\link{sprel16}}
}
\concept{Core}
\concept{Religion}
\keyword{variable}
