% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_158.R
\name{sphedindstatus}
\alias{sphedindstatus}
\title{Sphedind10 coding status}
\source{
General Social Survey https://gss.norc.org
}
\description{
sphedindstatus
}
\details{
Question SPHEDIND10 Coding Status
}
\section{Values}{

\itemize{
\item \code{1} found, iap
\item \code{2} found, no answer
\item \code{3} found, coded
\item \code{4} not found, iap
\item \code{5} not found, no answer
\item \code{6} not found, coded
\item \code{7} not found, no code assigned
\item \code{8} case not collected in 1994
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6006/vshow}{GSS Data Explorer page for \code{sphedindstatus}}.
}

\seealso{
Other Family Mobility: 
\code{\link{sphedind}},
\code{\link{sphedind10}},
\code{\link{sphedindtag}},
\code{\link{sphedocc10}},
\code{\link{sphedoccindv}},
\code{\link{sphedoccstatus}},
\code{\link{sphedocctag}},
\code{\link{sphedpre10}},
\code{\link{sphedslf}}
}
\concept{Family Mobility}
\keyword{variable}
