% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_158.R
\name{sphedocctag}
\alias{sphedocctag}
\title{Sphedocc10 coding tag}
\source{
General Social Survey https://gss.norc.org
}
\description{
sphedocctag
}
\details{
Question SPHEDOCC10 Coding Tag
}
\section{Values}{

\itemize{
\item \code{1} found
\item \code{2} not found, assigned
\item \code{3} not found, not assigned
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6004/vshow}{GSS Data Explorer page for \code{sphedocctag}}.
}

\seealso{
Other Family Mobility: 
\code{\link{sphedind}},
\code{\link{sphedind10}},
\code{\link{sphedindstatus}},
\code{\link{sphedindtag}},
\code{\link{sphedocc10}},
\code{\link{sphedoccindv}},
\code{\link{sphedoccstatus}},
\code{\link{sphedpre10}},
\code{\link{sphedslf}}
}
\concept{Family Mobility}
\keyword{variable}
