% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_158.R
\name{sphedpre10}
\alias{sphedpre10}
\title{Spouse's father's occ pres when spouse was 16 (2010)}
\source{
General Social Survey https://gss.norc.org
}
\description{
sphedpre10
}
\details{
Question Spouse's father's prestige score when industry was 16. SPHEDPRE10 is based on the 2010 Census occupation classification. This standard prestige score is a simple mean value of ratings for each occupation category, converted to a scale of 0 (bottom) to 100 (top).
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6008/vshow}{GSS Data Explorer page for \code{sphedpre10}}.
}

\seealso{
Other Family Mobility: 
\code{\link{sphedind}},
\code{\link{sphedind10}},
\code{\link{sphedindstatus}},
\code{\link{sphedindtag}},
\code{\link{sphedocc10}},
\code{\link{sphedoccindv}},
\code{\link{sphedoccstatus}},
\code{\link{sphedocctag}},
\code{\link{sphedslf}}
}
\concept{Family Mobility}
\keyword{variable}
