% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_159.R
\name{sphedpre105plus}
\alias{sphedpre105plus}
\title{Spouse's father's occ pres when spouse was 16 using threshold method (2010)}
\source{
General Social Survey https://gss.norc.org
}
\description{
sphedpre105plus
}
\details{
Question SPHEDPRE105PLUS is based on the 2010 Census occupation classification. This prestige score is calculated using an alternative method, based on the percentage of ratings that is greater than or equal to a threshold (rating five). The effect of individual raters is also removed using hierarchical linear modeling (HLM).
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6009/vshow}{GSS Data Explorer page for \code{sphedpre105plus}}.
}

\seealso{
Other Family Mobility: 
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{spsibs}}
}
\concept{Family Mobility}
\keyword{variable}
