% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_158.R
\name{sphedslf}
\alias{sphedslf}
\title{Spouse's father self-emp when spouse was 16?}
\source{
General Social Survey https://gss.norc.org
}
\description{
sphedslf
}
\details{
Question A. Now we would like to find out what kind of work your (husband's/wife's) father (OR SUBSTITUTE) did when your (husband/wife) was about 16 years old. What kind of work was your (husband's/wife's) father (OR SUBSTITUTE) doing? That is, what was (his/her) job called?
B. IF NOT ALREADY ANSWERED, ASK:  What did (he/she) actually do in that job?  Tell me what were some of  (his/her) main duties?
C. What kind of place did (he/she) work for?
D. IF NOT ALREADY ANSWERED, ASK:  What did they (make/do)?
E. IF ALREADY ANSWERED, CODE WITHOUT ASKING:  Was (he/she) self-employed or did (he/she) work for someone else?
}
\section{Values}{

\itemize{
\item \code{1} self employed
\item \code{2} someone else
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1489/vshow}{GSS Data Explorer page for \code{sphedslf}}.
}

\seealso{
Other Family Mobility: 
\code{\link{sphedind}},
\code{\link{sphedind10}},
\code{\link{sphedindstatus}},
\code{\link{sphedindtag}},
\code{\link{sphedocc10}},
\code{\link{sphedoccindv}},
\code{\link{sphedoccstatus}},
\code{\link{sphedocctag}},
\code{\link{sphedpre10}}
}
\concept{Family Mobility}
\keyword{variable}
