% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_177.R
\name{sphmewrk}
\alias{sphmewrk}
\title{How much housework r's spouse does}
\source{
General Social Survey https://gss.norc.org
}
\description{
sphmewrk
}
\details{
Question We're interested in knowing how much of the work around your home is done by different people. We don't mean taking care of children - just regular work around the house, like cooking, grocery shopping and doing little repair jobs.
B. How much does your spouse or partner do?
}
\section{Values}{

\itemize{
\item \verb{[1]} all
\item \verb{[2]} most
\item \verb{[3]} about half
\item \verb{[4]} some
\item \verb{[5]} very little or none
\item \verb{[6]} not married-living with partner
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1661/vshow}{GSS Data Explorer page for \code{sphmewrk}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab all \tab most \tab about half \tab some \tab very little or none \tab no answer \tab Total \cr
   1996 \tab 2124 \tab 19 \tab 108 \tab 355 \tab 181 \tab 100 \tab 17 \tab 2904 \cr
   Total \tab 2124 \tab 19 \tab 108 \tab 355 \tab 181 \tab 100 \tab 17 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{emoteoth}},
\code{\link{fejobaff}},
\code{\link{hapboys}},
\code{\link{hapgirls}},
\code{\link{rhmewrk}},
\code{\link{sharesep}},
\code{\link{tradmod}},
\code{\link{willwed1}},
\code{\link{willwed2}}
}
\concept{Gender}
\keyword{variable}
