% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_04.R
\name{sphrs2}
\alias{sphrs2}
\title{No. of hrs spouse usually works a week}
\source{
General Social Survey https://gss.norc.org
}
\description{
sphrs2
}
\details{
Question Last week was your (wife/husband) working full time, part time, going to school, keeping house, or what?
B. IF WITH A JOB, BUT NOT AT WORK: How many hours a week does (he/she) usually work, at all jobs?
}
\section{Values}{

\itemize{
\item \code{89} 89+ hours
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\seealso{
Other Core: 
\code{\link{coevwork}},
\code{\link{cohrs1}},
\code{\link{cowrkslf}},
\code{\link{cowrksta}},
\code{\link{divorce}},
\code{\link{spevwork}},
\code{\link{sphrs1}},
\code{\link{spwrksta}},
\code{\link{widowed}}
}
\concept{Core}
\keyword{variable}
