% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_37.R
\name{spirits}
\alias{spirits}
\title{Felt in touch with deceased person}
\source{
General Social Survey https://gss.norc.org
}
\description{
spirits
}
\details{
Question How often have you had any of the following experiences?  READ EACH ITEM AND CIRCLE ONE CODE FOR EACH.
D. Felt as though you were really in touch with someone who had died.
}
\section{Values}{

\itemize{
\item \verb{[1]} never in my life
\item \verb{[2]} once or twice
\item \verb{[3]} several times
\item \verb{[4]} often
\item \verb{[8]} cannot answer
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1989 \cr
   A/B/C \tab 1988, 1991 \cr
   Full \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/319/vshow}{GSS Data Explorer page for \code{spirits}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab never in my life \tab once or twice \tab several times \tab often \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 834 \tab 337 \tab 200 \tab 74 \tab 9 \tab 19 \tab 1473 \cr
   1988 \tab - \tab 877 \tab 353 \tab 148 \tab 81 \tab 18 \tab 4 \tab 1481 \cr
   1989 \tab 531 \tab 638 \tab 228 \tab 82 \tab 43 \tab 10 \tab 5 \tab 1537 \cr
   1991 \tab 158 \tab 708 \tab 283 \tab 135 \tab 47 \tab 165 \tab 21 \tab 1517 \cr
   Total \tab 689 \tab 3057 \tab 1201 \tab 565 \tab 245 \tab 202 \tab 49 \tab 6008 \cr
}
}

\seealso{
Other Core: 
\code{\link{dejavu}},
\code{\link{esp}},
\code{\link{grace}},
\code{\link{likediff}},
\code{\link{mindbody}},
\code{\link{palefull}},
\code{\link{pray}},
\code{\link{restact}},
\code{\link{visions}}
}
\concept{Core}
\keyword{variable}
