% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_37.R
\name{spirits}
\alias{spirits}
\title{Felt in touch with deceased person}
\source{
General Social Survey https://gss.norc.org
}
\description{
spirits
}
\details{
Question How often have you had any of the following experiences?  READ EACH ITEM AND CIRCLE ONE CODE FOR EACH.
D. Felt as though you were really in touch with someone who had died.
}
\section{Values}{

\itemize{
\item \code{1} never in my life
\item \code{2} once or twice
\item \code{3} several times
\item \code{4} often
\item \code{8} cannot answer
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab -/-/- \tab full \cr
   1988 \tab A/B/C \tab full \cr
   1989 \tab A/B/- \tab partial \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/319/vshow}{GSS Data Explorer page for \code{spirits}}.
}

\seealso{
Other Core: 
\code{\link{dejavu}},
\code{\link{esp}},
\code{\link{grace}},
\code{\link{likediff}},
\code{\link{mindbody}},
\code{\link{palefull}},
\code{\link{pray}},
\code{\link{restact}},
\code{\link{visions}}
}
\concept{Core}
\keyword{variable}
