% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_418.R
\name{spjoth16}
\alias{spjoth16}
\title{R's spouse raised jewish}
\source{
General Social Survey https://gss.norc.org
}
\description{
spjoth16
}
\details{
Question Was your (husband/wife) raised Jewish?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} half/part
\item \verb{[3]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3825/vshow}{GSS Data Explorer page for \code{spjoth16}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no \tab Total \cr
   2008 \tab 2020 \tab 1 \tab 2 \tab 2023 \cr
   2012 \tab 1971 \tab 1 \tab 2 \tab 1974 \cr
   Total \tab 3991 \tab 2 \tab 4 \tab 3997 \cr
}
}

\seealso{
Other Jewish Identity: 
\code{\link{majwoth}},
\code{\link{pajwoth}},
\code{\link{spjrel16}}
}
\concept{Jewish Identity}
\keyword{variable}
