% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_418.R
\name{spjrel16}
\alias{spjrel16}
\title{Religion in which spouse raised}
\source{
General Social Survey https://gss.norc.org
}
\description{
spjrel16
}
\details{
Question Religion in which spouse was raised.
}
\section{Values}{

\itemize{
\item \code{1} protestant
\item \code{2} catholic
\item \code{3} jewish
\item \code{4} none
\item \code{5} other
\item \code{6} buddhism
\item \code{7} hinduism
\item \code{8} other eastern religions
\item \code{9} muslim/islam
\item \code{10} orthodox-christian
\item \code{11} christian
\item \code{12} native american
\item \code{13} inter-nondenominational
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3824/vshow}{GSS Data Explorer page for \code{spjrel16}}.
}

\seealso{
Other Jewish Identity: 
\code{\link{majwoth}},
\code{\link{pajwoth}},
\code{\link{spjoth16}}
}
\concept{Jewish Identity}
\keyword{variable}
