% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_30.R
\name{spkath}
\alias{spkath}
\title{Allow anti-religionist to speak}
\source{
General Social Survey https://gss.norc.org
}
\description{
spkath
}
\details{
Question There are always some people whose ideas are considered bad or dangerous by other people. For instance, somebody who is against all churches and religion . . .
A. If such a person wanted to make a speech in your (city/town/community) against churches and religion, should he be allowed to speak, or not?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, allowed to speak
\item \verb{[2]} not allowed
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1972, 1973, 1974, 1976, 1977, 1980, 1982, 1984, 1985, 1987 \cr
   A/C \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/246/vshow}{GSS Data Explorer page for \code{spkath}}.

Counts by year:\tabular{llllllll}{
   year \tab yes, allowed to speak \tab not allowed \tab don't know \tab no answer \tab iap \tab skipped on web \tab Total \cr
   1972 \tab 1049 \tab 527 \tab 37 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 982 \tab 511 \tab 10 \tab 1 \tab - \tab - \tab 1504 \cr
   1974 \tab 915 \tab 547 \tab 20 \tab 2 \tab - \tab - \tab 1484 \cr
   1976 \tab 959 \tab 524 \tab 16 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 952 \tab 566 \tab 9 \tab 3 \tab - \tab - \tab 1530 \cr
   1980 \tab 971 \tab 485 \tab 12 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1140 \tab 684 \tab 35 \tab 1 \tab - \tab - \tab 1860 \cr
   1984 \tab 1002 \tab 459 \tab 8 \tab 4 \tab - \tab - \tab 1473 \cr
   1985 \tab 991 \tab 519 \tab 22 \tab 2 \tab - \tab - \tab 1534 \cr
   1987 \tab 1236 \tab 564 \tab 13 \tab 6 \tab - \tab - \tab 1819 \cr
   1988 \tab 681 \tab 280 \tab 15 \tab 1 \tab 504 \tab - \tab 1481 \cr
   1989 \tab 740 \tab 280 \tab 11 \tab 2 \tab 504 \tab - \tab 1537 \cr
   1990 \tab 663 \tab 236 \tab 15 \tab 3 \tab 455 \tab - \tab 1372 \cr
   1991 \tab 713 \tab 266 \tab 11 \tab 3 \tab 524 \tab - \tab 1517 \cr
   1993 \tab 764 \tab 290 \tab 19 \tab 2 \tab 531 \tab - \tab 1606 \cr
   1994 \tab 1451 \tab 520 \tab 21 \tab 4 \tab 996 \tab - \tab 2992 \cr
   1996 \tab 1409 \tab 483 \tab 27 \tab 4 \tab 981 \tab - \tab 2904 \cr
   1998 \tab 1393 \tab 457 \tab 29 \tab 3 \tab 950 \tab - \tab 2832 \cr
   2000 \tab 1381 \tab 450 \tab 25 \tab 5 \tab 956 \tab - \tab 2817 \cr
   2002 \tab 710 \tab 206 \tab 8 \tab - \tab 1841 \tab - \tab 2765 \cr
   2004 \tab 689 \tab 200 \tab 7 \tab 2 \tab 1914 \tab - \tab 2812 \cr
   2006 \tab 1553 \tab 438 \tab 12 \tab - \tab 2507 \tab - \tab 4510 \cr
   2008 \tab 1021 \tab 315 \tab 11 \tab 5 \tab 671 \tab - \tab 2023 \cr
   2010 \tab 956 \tab 312 \tab 11 \tab 2 \tab 763 \tab - \tab 2044 \cr
   2012 \tab 1011 \tab 281 \tab 15 \tab 1 \tab 666 \tab - \tab 1974 \cr
   2014 \tab 1345 \tab 351 \tab 12 \tab 6 \tab 824 \tab - \tab 2538 \cr
   2016 \tab 1493 \tab 378 \tab 11 \tab 8 \tab 977 \tab - \tab 2867 \cr
   2018 \tab 1253 \tab 309 \tab 8 \tab 4 \tab 774 \tab - \tab 2348 \cr
   2021 \tab 1068 \tab 244 \tab 1 \tab - \tab 2714 \tab 5 \tab 4032 \cr
   2022 \tab 909 \tab 238 \tab 11 \tab 1 \tab 2380 \tab 5 \tab 3544 \cr
   Total \tab 31400 \tab 11920 \tab 462 \tab 75 \tab 22432 \tab 10 \tab 66299 \cr
}
}

\seealso{
Other Core: 
\code{\link{colath}},
\code{\link{colsoc}},
\code{\link{educop}},
\code{\link{eqwlth}},
\code{\link{eqwlthy}},
\code{\link{govcare}},
\code{\link{libath}},
\code{\link{spksoc}},
\code{\link{tax}}
}
\concept{Core}
\keyword{variable}
