% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_623.R
\name{spkathy}
\alias{spkathy}
\title{Allow anti-religionist to speak y}
\source{
General Social Survey https://gss.norc.org
}
\description{
spkathy
}
\details{
Question Or consider a person who believes that Blacks are genetically inferior… If such a person wanted to make a speech in your community claiming that Blacks are inferior, should this person be allowed to speak, or not?
}
\section{Values}{

\itemize{
\item \code{1} yes, allowed to speak
\item \code{2} not allowed
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/C \tab partial \cr
   2022 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7434/vshow}{GSS Data Explorer page for \code{spkathy}}.
}

\keyword{variable}
