% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_253.R
\name{spklang}
\alias{spklang}
\title{How well does r speak other language}
\source{
General Social Survey https://gss.norc.org
}
\description{
spklang
}
\details{
Question E. How well do you speak that language? (IF SPEAKS 2 OR MORE, ASK ONLY OF THE MOST FLUENT LANGUAGE) READ CATEGORIES
}
\section{Values}{

\itemize{
\item \code{1} very well
\item \code{2} well
\item \code{3} not well
\item \code{4} poorly/hardly at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2287/vshow}{GSS Data Explorer page for \code{spklang}}.
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{contjpn}},
\code{\link{contmslm}},
\code{\link{getlang}},
\code{\link{othbest}},
\code{\link{othlang}},
\code{\link{othlang1}},
\code{\link{othlang2}},
\code{\link{parlang}},
\code{\link{uselang}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
