% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_534.R
\name{spklang2}
\alias{spklang2}
\title{What language(s) do you speak well?}
\source{
General Social Survey https://gss.norc.org
}
\description{
spklang2
}
\details{
Question What language(s) do you speak well?  PLEASE FILL IN.
Second response
}
\section{Values}{

\itemize{
\item \code{1} english
\item \code{2} spanish
\item \code{3} polish
\item \code{4} french
\item \code{5} hebrew
\item \code{6} russian
\item \code{7} tagalog
\item \code{8} chinese
\item \code{9} urdu
\item \code{10} italian
\item \code{11} romanian
\item \code{12} german
\item \code{13} pashtu
\item \code{14} chamorro
\item \code{15} ukranian
\item \code{16} hungarian
\item \code{17} turkish
\item \code{18} khmer
\item \code{19} korean
\item \code{20} navajo
\item \code{21} malayam
\item \code{22} arabic
\item \code{23} farsi-persian
\item \code{24} serbian
\item \code{25} lithuanian
\item \code{27} thai
\item \code{28} gujarati
\item \code{30} dari
\item \code{31} slovak
\item \code{32} japanese
\item \code{33} portuguese
\item \code{34} vajun
\item \code{35} yidddish
\item \code{36} creole
\item \code{37} vietnamese
\item \code{38} norwegian
\item \code{39} malaysian
\item \code{40} latin
\item \code{41} czech
\item \code{42} hindu
\item \code{43} hakka
\item \code{44} swedish
\item \code{45} vesaya
\item \code{46} armenian
\item \code{47} american sign language
\item \code{48} cambodian
\item \code{49} spanglish
\item \code{50} tigrigna
\item \code{51} oriya
\item \code{52} hnong
\item \code{53} marathi
\item \code{54} wolof
\item \code{55} pangasian
\item \code{56} scottish
\item \code{57} macednian
\item \code{58} hawaiian
\item \code{59} fulani
\item \code{60} greek
\item \code{61} cajun
\item \code{62} akkadian
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4871/vshow}{GSS Data Explorer page for \code{spklang2}}.
}

\seealso{
Other ISSP National Identity: 
\code{\link{abroad}},
\code{\link{amancstr}},
\code{\link{ambetter}},
\code{\link{ambornin}},
\code{\link{amchrstn}},
\code{\link{amcit}},
\code{\link{amcitizn}},
\code{\link{amcult}},
\code{\link{amenglsh}},
\code{\link{amfeel}},
\code{\link{amgovt}},
\code{\link{amlived}},
\code{\link{amownway}},
\code{\link{amproud1}},
\code{\link{amshamed}},
\code{\link{amsports}},
\code{\link{amtv}},
\code{\link{belikeus}},
\code{\link{biasnews}},
\code{\link{bizmoney}},
\code{\link{brneffrt}},
\code{\link{citizen}},
\code{\link{citworld}},
\code{\link{clsenei}},
\code{\link{clsenoam}},
\code{\link{clsestat}},
\code{\link{clsetown}},
\code{\link{clseusa}},
\code{\link{conpow}},
\code{\link{conscrp}},
\code{\link{decsorgs}},
\code{\link{ethclose}},
\code{\link{experts}},
\code{\link{forcult}},
\code{\link{forland}},
\code{\link{forlang}},
\code{\link{freetrde}},
\code{\link{fulldem}},
\code{\link{ifwrong}},
\code{\link{immameco}},
\code{\link{immassim}},
\code{\link{immcosts}},
\code{\link{immcrime1}},
\code{\link{immcult}},
\code{\link{immeduc}},
\code{\link{immideas}},
\code{\link{immideas1}},
\code{\link{immimp}},
\code{\link{immjobs1}},
\code{\link{immrghts}},
\code{\link{immwlfare}},
\code{\link{imports}},
\code{\link{incgap1}},
\code{\link{internet}},
\code{\link{intlincs}},
\code{\link{kidsaway}},
\code{\link{kidshere}},
\code{\link{lessprd}},
\code{\link{letin1}},
\code{\link{letin1a}},
\code{\link{livecom1}},
\code{\link{meltpot1}},
\code{\link{mincult}},
\code{\link{movenei}},
\code{\link{movenoam}},
\code{\link{movestat}},
\code{\link{movetown}},
\code{\link{moveusa}},
\code{\link{nafta1}},
\code{\link{nafta2}},
\code{\link{nafta2a}},
\code{\link{nafta3}},
\code{\link{onenatn}},
\code{\link{parcit}},
\code{\link{patriot1}},
\code{\link{patriot2}},
\code{\link{patriot3}},
\code{\link{patriot4}},
\code{\link{powrorgs}},
\code{\link{proudart}},
\code{\link{prouddem}},
\code{\link{proudeco}},
\code{\link{proudgrp}},
\code{\link{proudhis}},
\code{\link{proudmil}},
\code{\link{proudpol}},
\code{\link{proudsci}},
\code{\link{proudspt}},
\code{\link{proudsss}},
\code{\link{refugees}},
\code{\link{res161}},
\code{\link{shortcom}},
\code{\link{socid1}},
\code{\link{socid2}},
\code{\link{socid3}},
\code{\link{spkhome1}},
\code{\link{spkhome2}},
\code{\link{spklang1}},
\code{\link{spklang3}},
\code{\link{wrldgovt}}
}
\concept{ISSP National Identity}
\keyword{variable}
