% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_255.R
\name{spklangw}
\alias{spklangw}
\title{How well r speaks the language in betrlang}
\source{
General Social Survey https://gss.norc.org
}
\description{
spklangw
}
\details{
Question C. How well do you speak that language?
}
\section{Values}{

\itemize{
\item \code{1} very well
\item \code{2} well
\item \code{3} not well
\item \code{4} poorly/hardly at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2303/vshow}{GSS Data Explorer page for \code{spklangw}}.
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{betrlang}},
\code{\link{immcrmup}},
\code{\link{immnew}},
\code{\link{immnojob}},
\code{\link{letinasn}},
\code{\link{letineur}},
\code{\link{letinhsp}},
\code{\link{othhome}},
\code{\link{uswht}}

Other Language: 
\code{\link{betrlang}},
\code{\link{othhome}}
}
\concept{Language}
\concept{Multi-Ethnic United States}
\keyword{variable}
