% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_624.R
\name{spkmily}
\alias{spkmily}
\title{Allow militarist to speak y}
\source{
General Social Survey https://gss.norc.org
}
\description{
spkmily
}
\details{
Question Consider a person who advocates doing away with elections and letting the military run the country. If such a person wanted to make a speech in your community, should this person be allowed to speak, or not?
}
\section{Values}{

\itemize{
\item \code{1} yes, allowed to speak
\item \code{2} not allowed
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7441/vshow}{GSS Data Explorer page for \code{spkmily}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes, allowed to speak \tab not allowed \tab don't know \tab skipped on web \tab Total \cr
   2021 \tab 2675 \tab 839 \tab 504 \tab 3 \tab 11 \tab 4032 \cr
   Total \tab 2675 \tab 839 \tab 504 \tab 3 \tab 11 \tab 4032 \cr
}
}

\keyword{variable}
