% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_491.R
\name{splive}
\alias{splive}
\title{How long been living with spouse}
\source{
General Social Survey https://gss.norc.org
}
\description{
splive
}
\details{
Question How long have you and your spouse/steady partner been living together?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4485/vshow}{GSS Data Explorer page for \code{splive}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{rfamlook}},
\code{\link{spfalook}},
\code{\link{wkyngscs}}
}
\concept{Cohabitation}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
