% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_491.R
\name{splive}
\alias{splive}
\title{How long been living with spouse}
\source{
General Social Survey https://gss.norc.org
}
\description{
splive
}
\details{
Question How long have you and your spouse/steady partner been living together?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4485/vshow}{GSS Data Explorer page for \code{splive}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 3 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 4 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 5 \tab 50 \tab 51 \tab 52 \tab 53 \tab 55 \tab 56 \tab 57 \tab 58 \tab 59 \tab 6 \tab 61 \tab 63 \tab 64 \tab 68 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2012 \tab 1399 \tab 1 \tab 5 \tab 18 \tab 13 \tab 25 \tab 23 \tab 10 \tab 18 \tab 12 \tab 10 \tab 14 \tab 10 \tab 13 \tab 17 \tab 10 \tab 9 \tab 7 \tab 6 \tab 11 \tab 13 \tab 9 \tab 9 \tab 5 \tab 15 \tab 14 \tab 6 \tab 7 \tab 9 \tab 6 \tab 8 \tab 7 \tab 5 \tab 4 \tab 7 \tab 18 \tab 5 \tab 6 \tab 5 \tab 3 \tab 3 \tab 2 \tab 5 \tab 5 \tab 6 \tab 3 \tab 20 \tab 7 \tab 4 \tab 5 \tab 5 \tab 3 \tab 5 \tab 1 \tab 3 \tab 3 \tab 22 \tab 1 \tab 1 \tab 2 \tab 1 \tab 25 \tab 24 \tab 15 \tab 2 \tab 9 \tab 1974 \cr
   Total \tab 1399 \tab 1 \tab 5 \tab 18 \tab 13 \tab 25 \tab 23 \tab 10 \tab 18 \tab 12 \tab 10 \tab 14 \tab 10 \tab 13 \tab 17 \tab 10 \tab 9 \tab 7 \tab 6 \tab 11 \tab 13 \tab 9 \tab 9 \tab 5 \tab 15 \tab 14 \tab 6 \tab 7 \tab 9 \tab 6 \tab 8 \tab 7 \tab 5 \tab 4 \tab 7 \tab 18 \tab 5 \tab 6 \tab 5 \tab 3 \tab 3 \tab 2 \tab 5 \tab 5 \tab 6 \tab 3 \tab 20 \tab 7 \tab 4 \tab 5 \tab 5 \tab 3 \tab 5 \tab 1 \tab 3 \tab 3 \tab 22 \tab 1 \tab 1 \tab 2 \tab 1 \tab 25 \tab 24 \tab 15 \tab 2 \tab 9 \tab 1974 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{rfamlook}},
\code{\link{spfalook}},
\code{\link{wkyngscs}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
