% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_146.R
\name{splocnum}
\alias{splocnum}
\title{#Of workers at the location your spouse works}
\source{
General Social Survey https://gss.norc.org
}
\description{
splocnum
}
\details{
Question About how may people work at the location where your (husband/wife) works? (HAND CARD BY IF NOT ALREADY GIVEN TO R.) Count part-time as well as full-time employees in all areas, departments and buildings.
}
\section{Values}{

\itemize{
\item \verb{[1]} 1-9
\item \verb{[2]} 10-49
\item \verb{[3]} 50-99
\item \verb{[4]} 100-499
\item \verb{[5]} 500-999
\item \verb{[6]} 1,000-1,999
\item \verb{[7]} 2,000+
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1375/vshow}{GSS Data Explorer page for \code{splocnum}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab 1-9 \tab 10-49 \tab 50-99 \tab 100-499 \tab 500-999 \tab 1,000-1,999 \tab 2,000+ \tab don't know \tab no answer \tab Total \cr
   1991 \tab 1001 \tab 126 \tab 110 \tab 56 \tab 100 \tab 20 \tab 19 \tab 41 \tab 30 \tab 14 \tab 1517 \cr
   Total \tab 1001 \tab 126 \tab 110 \tab 56 \tab 100 \tab 20 \tab 19 \tab 41 \tab 30 \tab 14 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{baseofop}},
\code{\link{exptrain}},
\code{\link{findout}},
\code{\link{learning}},
\code{\link{training}},
\code{\link{wherewrk}}
}
\concept{Work Organizations}
\keyword{variable}
