% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_146.R
\name{splocnum}
\alias{splocnum}
\title{#Of workers at the location your spouse works}
\source{
General Social Survey https://gss.norc.org
}
\description{
splocnum
}
\details{
Question About how may people work at the location where your (husband/wife) works? (HAND CARD BY IF NOT ALREADY GIVEN TO R.) Count part-time as well as full-time employees in all areas, departments and buildings.
}
\section{Values}{

\itemize{
\item \code{1} 1-9
\item \code{2} 10-49
\item \code{3} 50-99
\item \code{4} 100-499
\item \code{5} 500-999
\item \code{6} 1,000-1,999
\item \code{7} 2,000+
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1375/vshow}{GSS Data Explorer page for \code{splocnum}}.
}

\seealso{
Other Work Organizations: 
\code{\link{baseofop}},
\code{\link{exptrain}},
\code{\link{findout}},
\code{\link{learning}},
\code{\link{training}},
\code{\link{wherewrk}}

Other Job: 
\code{\link{baseofop}},
\code{\link{exptrain}},
\code{\link{findout}},
\code{\link{learning}},
\code{\link{training}},
\code{\link{wherewrk}}
}
\concept{Job}
\concept{Work Organizations}
\keyword{variable}
