% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_157.R
\name{spmaeduc}
\alias{spmaeduc}
\title{Highest schl grade spouse's mother completed}
\source{
General Social Survey https://gss.norc.org
}
\description{
spmaeduc
}
\details{
Question B. What is the highest grade or year of regular school that your (husband's/wife's) mother ever completed?
CODE EXACT GRADE.
}
\section{Values}{

\itemize{
\item \verb{[0]} no formal schooling
\item \verb{[1]} 1st grade
\item \verb{[2]} 2nd grade
\item \verb{[3]} 3rd grade
\item \verb{[4]} 4th grade
\item \verb{[5]} 5th grade
\item \verb{[6]} 6th grade
\item \verb{[7]} 7th grade
\item \verb{[8]} 8th grade
\item \verb{[9]} 9th grade
\item \verb{[10]} 10th grade
\item \verb{[11]} 11th grade
\item \verb{[12]} 12th grade
\item \verb{[13]} 1 year of college
\item \verb{[14]} 2 years of college
\item \verb{[15]} 3 years of college
\item \verb{[16]} 4 years of college
\item \verb{[17]} 5 years of college
\item \verb{[18]} 6 years of college
\item \verb{[19]} 7 years of college
\item \verb{[20]} 8 or more years of college
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1486/vshow}{GSS Data Explorer page for \code{spmaeduc}}.

Counts by year:\tabular{llllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 2 \tab 20 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2247 \tab 6 \tab 15 \tab 6 \tab 255 \tab 8 \tab 33 \tab 5 \tab 45 \tab 2 \tab 9 \tab 1 \tab 2 \tab 5 \tab 4 \tab 7 \tab 10 \tab 4 \tab 60 \tab 17 \tab 222 \tab 29 \tab 2992 \cr
   Total \tab 2247 \tab 6 \tab 15 \tab 6 \tab 255 \tab 8 \tab 33 \tab 5 \tab 45 \tab 2 \tab 9 \tab 1 \tab 2 \tab 5 \tab 4 \tab 7 \tab 10 \tab 4 \tab 60 \tab 17 \tab 222 \tab 29 \tab 2992 \cr
}
}

\seealso{
Other Family Mobility: 
\code{\link{agewedcr}},
\code{\link{fstspmed}},
\code{\link{fstspped}},
\code{\link{spfam16}},
\code{\link{sphedocc}},
\code{\link{sphedpre}},
\code{\link{spmarnum}},
\code{\link{sppaeduc}},
\code{\link{spyrborn}}
}
\concept{Family Mobility}
\keyword{variable}
