% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_157.R
\name{spmarnum}
\alias{spmarnum}
\title{Has current spouse been married before?}
\source{
General Social Survey https://gss.norc.org
}
\description{
spmarnum
}
\details{
Question Has your (husband/wife) been married more than once?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1483/vshow}{GSS Data Explorer page for \code{spmarnum}}.
}

\seealso{
Other Family Mobility: 
\code{\link{agewedcr}},
\code{\link{fstspmed}},
\code{\link{fstspped}},
\code{\link{spfam16}},
\code{\link{sphedocc}},
\code{\link{sphedpre}},
\code{\link{spmaeduc}},
\code{\link{sppaeduc}},
\code{\link{spyrborn}}

Other Marital Status: 
\code{\link{agewedcr}}
}
\concept{Family Mobility}
\concept{Marital Status}
\keyword{variable}
