% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_215.R
\name{spmentl}
\alias{spmentl}
\title{Govt spending on health}
\source{
General Social Survey https://gss.norc.org
}
\description{
spmentl
}
\details{
Question Earlier, we talked about various areas of government spending. Since we've been talking about the mental health area, please indicate whether you would like to see more or less government spending in the area of mental health care. Remember that if you say "much more," it might require a tax increase to pay for it.
}
\section{Values}{

\itemize{
\item \verb{[1]} spend much more
\item \verb{[2]} spend more
\item \verb{[3]} spend the same as now
\item \verb{[4]} spend less
\item \verb{[5]} spend much less
\item \verb{[8]} cant choose
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2038/vshow}{GSS Data Explorer page for \code{spmentl}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab spend much more \tab spend more \tab spend the same as now \tab spend less \tab spend much less \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2168 \tab 86 \tab 248 \tab 267 \tab 61 \tab 15 \tab 48 \tab 11 \tab 2904 \cr
   2006 \tab 3775 \tab 79 \tab 276 \tab 299 \tab 36 \tab 14 \tab 24 \tab 7 \tab 4510 \cr
   Total \tab 5943 \tab 165 \tab 524 \tab 566 \tab 97 \tab 29 \tab 72 \tab 18 \tab 7414 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{dangroth}},
\code{\link{dangrslf}},
\code{\link{govmentl}},
\code{\link{medcare1}},
\code{\link{medcare2}},
\code{\link{mntlas1}},
\code{\link{mustdoc}},
\code{\link{musthosp}},
\code{\link{mustmed}}
}
\concept{Mental Health}
\keyword{variable}
