% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_215.R
\name{spmentl}
\alias{spmentl}
\title{Govt spending on health}
\source{
General Social Survey https://gss.norc.org
}
\description{
spmentl
}
\details{
Question Earlier, we talked about various areas of government spending. Since we've been talking about the mental health area, please indicate whether you would like to see more or less government spending in the area of mental health care. Remember that if you say "much more," it might require a tax increase to pay for it.
}
\section{Values}{

\itemize{
\item \code{1} spend much more
\item \code{2} spend more
\item \code{3} spend the same as now
\item \code{4} spend less
\item \code{5} spend much less
\item \code{8} cant choose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2038/vshow}{GSS Data Explorer page for \code{spmentl}}.
}

\seealso{
Other Mental Health: 
\code{\link{dangroth}},
\code{\link{dangrslf}},
\code{\link{govmentl}},
\code{\link{medcare1}},
\code{\link{medcare2}},
\code{\link{mntlas1}},
\code{\link{mustdoc}},
\code{\link{musthosp}},
\code{\link{mustmed}}

Other Mental Health: 
\code{\link{dangroth}},
\code{\link{dangrslf}},
\code{\link{govmentl}},
\code{\link{medcare1}},
\code{\link{medcare2}},
\code{\link{mntlas1}},
\code{\link{mustdoc}},
\code{\link{musthosp}},
\code{\link{mustmed}}
}
\concept{Federal Government}
\concept{Mental Health}
\concept{Spending}
\keyword{variable}
