% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_406.R
\name{spnatdis}
\alias{spnatdis}
\title{Should govt spend more or less on natural disasters}
\source{
General Social Survey https://gss.norc.org
}
\description{
spnatdis
}
\details{
Question Would you like to see more or less government spending on natural disasters. Remember that if you say "much more", it might require a tax increase to pay for it. Would you say...
}
\section{Values}{

\itemize{
\item \verb{[1]} spend much more
\item \verb{[2]} spend more
\item \verb{[3]} spend the same as now
\item \verb{[4]} spend less
\item \verb{[5]} spend much less
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3707/vshow}{GSS Data Explorer page for \code{spnatdis}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab spend much more \tab spend more \tab spend the same as now \tab spend less \tab spend much less \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2992 \tab 212 \tab 511 \tab 577 \tab 158 \tab 38 \tab 19 \tab 3 \tab 4510 \cr
   Total \tab 2992 \tab 212 \tab 511 \tab 577 \tab 158 \tab 38 \tab 19 \tab 3 \tab 4510 \cr
}
}

\seealso{
Other Misc I: 
\code{\link{natdisin}}
}
\concept{Misc I}
\keyword{variable}
