% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_406.R
\name{spnatdis}
\alias{spnatdis}
\title{Should govt spend more or less on natural disasters}
\source{
General Social Survey https://gss.norc.org
}
\description{
spnatdis
}
\details{
Question Would you like to see more or less government spending on natural disasters. Remember that if you say "much more", it might require a tax increase to pay for it. Would you say...
}
\section{Values}{

\itemize{
\item \code{1} spend much more
\item \code{2} spend more
\item \code{3} spend the same as now
\item \code{4} spend less
\item \code{5} spend much less
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3707/vshow}{GSS Data Explorer page for \code{spnatdis}}.
}

\seealso{
Other Misc I: 
\code{\link{natdisin}}

Other Federal Government: 
\code{\link{natdisin}}
}
\concept{Federal Government}
\concept{Misc I}
\concept{Spending}
\keyword{variable}
