% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_06.R
\name{spoccstatus}
\alias{spoccstatus}
\title{Spocc10 coding status}
\source{
General Social Survey https://gss.norc.org
}
\description{
spoccstatus
}
\details{
Question SPOCC10 Coding Status
}
\section{Values}{

\itemize{
\item \code{1} found, iap
\item \code{2} found, no answer
\item \code{3} found, coded
\item \code{4} not found, iap
\item \code{5} not found, no answer
\item \code{6} not found, coded
\item \code{7} not found, no code assigned
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5987/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllllll}{
   year \tab found, coded \tab found, iap \tab found, no answer \tab not found, coded \tab not found, iap \tab not found, no answer \tab not found, no code assigned \tab iap \tab not available in this year \tab Total \cr
   1972 \tab 442 \tab 255 \tab 5 \tab 476 \tab 325 \tab 5 \tab 105 \tab - \tab - \tab 1613 \cr
   1973 \tab 783 \tab 439 \tab 12 \tab 102 \tab 97 \tab 5 \tab 66 \tab - \tab - \tab 1504 \cr
   1974 \tab 742 \tab 428 \tab 47 \tab 135 \tab 95 \tab 2 \tab 35 \tab - \tab - \tab 1484 \cr
   1975 \tab 611 \tab 361 \tab 4 \tab 226 \tab 210 \tab 1 \tab 77 \tab - \tab - \tab 1490 \cr
   1976 \tab 736 \tab 488 \tab 6 \tab 117 \tab 124 \tab 1 \tab 27 \tab - \tab - \tab 1499 \cr
   1977 \tab 486 \tab 323 \tab 8 \tab 311 \tab 315 \tab - \tab 87 \tab - \tab - \tab 1530 \cr
   1978 \tab 872 \tab 639 \tab 7 \tab - \tab 14 \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 724 \tab 566 \tab 3 \tab 58 \tab 74 \tab - \tab 43 \tab - \tab - \tab 1468 \cr
   1982 \tab 832 \tab 855 \tab 9 \tab 66 \tab 84 \tab 2 \tab 12 \tab - \tab - \tab 1860 \cr
   1983 \tab 821 \tab 650 \tab 4 \tab 56 \tab 54 \tab - \tab 14 \tab - \tab - \tab 1599 \cr
   1984 \tab 774 \tab 677 \tab 7 \tab 4 \tab 10 \tab - \tab 1 \tab - \tab - \tab 1473 \cr
   1985 \tab 689 \tab 602 \tab 5 \tab 106 \tab 107 \tab - \tab 25 \tab - \tab - \tab 1534 \cr
   1986 \tab 751 \tab 666 \tab 8 \tab 7 \tab 32 \tab - \tab 6 \tab - \tab - \tab 1470 \cr
   1987 \tab 730 \tab 777 \tab 4 \tab 84 \tab 172 \tab 5 \tab 47 \tab - \tab - \tab 1819 \cr
   1988 \tab 724 \tab 728 \tab 8 \tab 6 \tab 14 \tab - \tab 1 \tab - \tab - \tab 1481 \cr
   1989 \tab 739 \tab 666 \tab 15 \tab 47 \tab 68 \tab - \tab 2 \tab - \tab - \tab 1537 \cr
   1990 \tab 675 \tab 672 \tab 11 \tab 2 \tab 12 \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 757 \tab 733 \tab 9 \tab 3 \tab 15 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 794 \tab 743 \tab 8 \tab 20 \tab 38 \tab - \tab 3 \tab - \tab - \tab 1606 \cr
   1994 \tab 1439 \tab 1491 \tab 8 \tab 20 \tab 30 \tab 2 \tab 2 \tab - \tab - \tab 2992 \cr
   1996 \tab 1318 \tab 1551 \tab 14 \tab 6 \tab 12 \tab - \tab 3 \tab - \tab - \tab 2904 \cr
   1998 \tab 1277 \tab 1505 \tab 19 \tab 6 \tab 22 \tab 1 \tab 2 \tab - \tab - \tab 2832 \cr
   2000 \tab 1169 \tab 1608 \tab 40 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1200 \tab 1547 \tab 18 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1364 \tab 1406 \tab 42 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2002 \tab 2465 \tab 43 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 928 \tab 1079 \tab 16 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 839 \tab 1187 \tab 18 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 25218 \tab 25107 \tab 398 \tab 1858 \tab 1924 \tab 24 \tab 558 \tab 13759 \tab 6853 \tab 75699 \cr
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/-/- \tab full \cr
   1973 \tab -/-/-/- \tab full \cr
   1974 \tab -/-/-/- \tab full \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
}
}

\seealso{
Other Core: 
\code{\link{coind10}},
\code{\link{coocc10}},
\code{\link{paocc16}},
\code{\link{spind10}},
\code{\link{spindstatus}},
\code{\link{spindtag}},
\code{\link{spocctag}},
\code{\link{sppres10}},
\code{\link{sppres105plus}}
}
\concept{Core}
\keyword{variable}
