% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_06.R
\name{spoccstatus}
\alias{spoccstatus}
\title{Spocc10 coding status}
\source{
General Social Survey https://gss.norc.org
}
\description{
spoccstatus
}
\details{
Question SPOCC10 Coding Status
}
\section{Values}{

\itemize{
\item \code{1} found, iap
\item \code{2} found, no answer
\item \code{3} found, coded
\item \code{4} not found, iap
\item \code{5} not found, no answer
\item \code{6} not found, coded
\item \code{7} not found, no code assigned
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/-/- \tab full \cr
   1973 \tab -/-/-/- \tab full \cr
   1974 \tab -/-/-/- \tab full \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5987/vshow}{GSS Data Explorer page for \code{spoccstatus}}.
}

\seealso{
Other Core: 
\code{\link{coind10}},
\code{\link{coocc10}},
\code{\link{paocc16}},
\code{\link{spind10}},
\code{\link{spindstatus}},
\code{\link{spindtag}},
\code{\link{spocctag}},
\code{\link{sppres10}},
\code{\link{sppres105plus}}
}
\concept{Core}
\keyword{variable}
