% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_06.R
\name{spocctag}
\alias{spocctag}
\title{Spocc10 coding tag}
\source{
General Social Survey https://gss.norc.org
}
\description{
spocctag
}
\details{
Question SPOCC10 Coding Tag
}
\section{Values}{

\itemize{
\item \code{1} found
\item \code{2} not found, assigned
\item \code{3} not found, not assigned
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/-/- \tab full \cr
   1973 \tab -/-/-/- \tab full \cr
   1974 \tab -/-/-/- \tab full \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5988/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllll}{
   year \tab found \tab not found, assigned \tab not found, not assigned \tab iap \tab not available in this year \tab Total \cr
   1972 \tab 702 \tab 806 \tab 105 \tab - \tab - \tab 1613 \cr
   1973 \tab 1234 \tab 204 \tab 66 \tab - \tab - \tab 1504 \cr
   1974 \tab 1217 \tab 232 \tab 35 \tab - \tab - \tab 1484 \cr
   1975 \tab 976 \tab 437 \tab 77 \tab - \tab - \tab 1490 \cr
   1976 \tab 1230 \tab 242 \tab 27 \tab - \tab - \tab 1499 \cr
   1977 \tab 817 \tab 626 \tab 87 \tab - \tab - \tab 1530 \cr
   1978 \tab 1518 \tab 14 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1293 \tab 132 \tab 43 \tab - \tab - \tab 1468 \cr
   1982 \tab 1696 \tab 152 \tab 12 \tab - \tab - \tab 1860 \cr
   1983 \tab 1475 \tab 110 \tab 14 \tab - \tab - \tab 1599 \cr
   1984 \tab 1458 \tab 14 \tab 1 \tab - \tab - \tab 1473 \cr
   1985 \tab 1296 \tab 213 \tab 25 \tab - \tab - \tab 1534 \cr
   1986 \tab 1425 \tab 39 \tab 6 \tab - \tab - \tab 1470 \cr
   1987 \tab 1511 \tab 261 \tab 47 \tab - \tab - \tab 1819 \cr
   1988 \tab 1460 \tab 20 \tab 1 \tab - \tab - \tab 1481 \cr
   1989 \tab 1420 \tab 115 \tab 2 \tab - \tab - \tab 1537 \cr
   1990 \tab 1358 \tab 14 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1499 \tab 18 \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1545 \tab 58 \tab 3 \tab - \tab - \tab 1606 \cr
   1994 \tab 2938 \tab 52 \tab 2 \tab - \tab - \tab 2992 \cr
   1996 \tab 2883 \tab 18 \tab 3 \tab - \tab - \tab 2904 \cr
   1998 \tab 2801 \tab 29 \tab 2 \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab - \tab - \tab 1974 \tab - \tab 1974 \cr
   2014 \tab - \tab - \tab - \tab 2538 \tab - \tab 2538 \cr
   2016 \tab - \tab - \tab - \tab 2867 \tab - \tab 2867 \cr
   2018 \tab - \tab - \tab - \tab 2348 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab 4032 \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 50723 \tab 3806 \tab 558 \tab 13759 \tab 6853 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{coind10}},
\code{\link{coocc10}},
\code{\link{paocc16}},
\code{\link{spind10}},
\code{\link{spindstatus}},
\code{\link{spindtag}},
\code{\link{spoccstatus}},
\code{\link{sppres10}},
\code{\link{sppres105plus}}
}
\concept{Core}
\keyword{variable}
