% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_41.R
\name{spothjew}
\alias{spothjew}
\title{Spouse considers self jewish}
\source{
General Social Survey https://gss.norc.org
}
\description{
spothjew
}
\details{
Question What is your (SPOUSE''S) religious preference?  Is it Protestant, Catholic, Jewish, some other religion, or no religion?
A. DOES YOUR SPOUSE CONSIDER HIMSELF/HERSELF SELF AS JEWISH
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} half/part
\item \code{3} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/351/vshow}{GSS Data Explorer page for \code{spothjew}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab no \tab no answer \tab yes \tab Total \cr
   2008 \tab 2008 \tab 3 \tab 12 \tab - \tab 2023 \cr
   2012 \tab 1971 \tab 2 \tab - \tab 1 \tab 1974 \cr
   Total \tab 3979 \tab 5 \tab 12 \tab 1 \tab 3997 \cr
}
}

\seealso{
Other Core: 
\code{\link{corel}},
\code{\link{spden16}},
\code{\link{spfund}},
\code{\link{spfund16}},
\code{\link{spjew}},
\code{\link{spjew16}},
\code{\link{spjewaj}},
\code{\link{spoth16}},
\code{\link{sprel16}}
}
\concept{Core}
\keyword{variable}
