% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_70.R
\name{sppaid}
\alias{sppaid}
\title{Regularly received alimony - chld supprt}
\source{
General Social Survey https://gss.norc.org
}
\description{
sppaid
}
\details{
Question Are you currently or have you ever been entitled to alimony or child support payments from a former (husband/wife)?
A. During the time you were entitled to alimony or child support, did you receive these payments on a regular basis, occasionally, or not at all?
}
\section{Values}{

\itemize{
\item \verb{[1]} regular basis
\item \verb{[2]} occasionally
\item \verb{[3]} not at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1986 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/630/vshow}{GSS Data Explorer page for \code{sppaid}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab regular basis \tab occasionally \tab not at all \tab no answer \tab Total \cr
   1986 \tab 1316 \tab 57 \tab 29 \tab 36 \tab 32 \tab 1470 \cr
   Total \tab 1316 \tab 57 \tab 29 \tab 36 \tab 32 \tab 1470 \cr
}
}

\seealso{
Other Core: 
\code{\link{divlaw}},
\code{\link{divlawy}},
\code{\link{homochng}},
\code{\link{homosex}},
\code{\link{porninf}},
\code{\link{premarsx}},
\code{\link{spdue}},
\code{\link{teensex}},
\code{\link{xmarsex}}
}
\concept{Core}
\keyword{variable}
