% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_442.R
\name{sppolice}
\alias{sppolice}
\title{Govt spending on law enforcement}
\source{
General Social Survey https://gss.norc.org
}
\description{
sppolice
}
\details{
Question Listed below are various areas of government spending. Please indicate whether you would like to see more or less government spending in each area. Remember that if you say "much more," it might require a tax increase to pay for it.
C. The police and law enforcement.
}
\section{Values}{

\itemize{
\item \code{1} spend much more
\item \code{2} spend more
\item \code{3} spend same
\item \code{4} spend less
\item \code{5} spend much less
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1990, 1996 \cr
   D/-/- \tab 2006 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4059/vshow}{GSS Data Explorer page for \code{sppolice}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab spend much more \tab spend more \tab spend same \tab spend less \tab spend much less \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 64 \tab 274 \tab 275 \tab 31 \tab 6 \tab 12 \tab 15 \tab 1534 \cr
   1990 \tab 155 \tab 131 \tab 509 \tab 442 \tab 50 \tab 15 \tab 32 \tab 38 \tab 1372 \cr
   1996 \tab 1572 \tab 155 \tab 586 \tab 447 \tab 74 \tab 15 \tab 19 \tab 36 \tab 2904 \cr
   2006 \tab 2992 \tab 245 \tab 593 \tab 544 \tab 100 \tab 20 \tab 15 \tab 1 \tab 4510 \cr
   2016 \tab 1477 \tab 150 \tab 545 \tab 537 \tab 117 \tab 21 \tab 9 \tab 11 \tab 2867 \cr
   Total \tab 7053 \tab 745 \tab 2507 \tab 2245 \tab 372 \tab 77 \tab 87 \tab 101 \tab 13187 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{bustax}},
\code{\link{inctax}},
\code{\link{infljobs}},
\code{\link{sparms}},
\code{\link{sparts}},
\code{\link{sphlth}},
\code{\link{spretire}},
\code{\link{spschool}},
\code{\link{spunemp}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
