% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_560.R
\name{sppoorkd}
\alias{sppoorkd}
\title{Spending for childcare for poor}
\source{
General Social Survey https://gss.norc.org
}
\description{
sppoorkd
}
\details{
Question Here are some other areas of government spending. Please indicate whether you would like to see more or less government spending in each area.
Remember that if you say "much more", it might require a tax increase to pay for it.
D. Child care for poor children…
}
\section{Values}{

\itemize{
\item \code{1} spend much more
\item \code{2} spend more
\item \code{3} spend same
\item \code{4} spend less
\item \code{5} spend much less
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5125/vshow}{GSS Data Explorer page for \code{sppoorkd}}.
}

\seealso{
Other Children: 
\code{\link{spdrugs}},
\code{\link{spdsabkd}},
\code{\link{spfoodkd}},
\code{\link{spheadst}},
\code{\link{sphlthkd}},
\code{\link{sphomekd}},
\code{\link{sppill}},
\code{\link{spwrkpar}}
}
\concept{Children}
\keyword{variable}
