% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_559.R
\name{sppregnt}
\alias{sppregnt}
\title{Spending for prenatal care}
\source{
General Social Survey https://gss.norc.org
}
\description{
sppregnt
}
\details{
Question Here are some other areas of government spending. Please indicate whether you would like to see more or less government spending in each area.
Remember that if you say "much more", it might require a tax increase to pay for it.
A. Prenatal care for pregnant mothers who can't afford it…
}
\section{Values}{

\itemize{
\item \code{1} spend much more
\item \code{2} spend more
\item \code{3} spend same
\item \code{4} spend less
\item \code{5} spend much less
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5122/vshow}{GSS Data Explorer page for \code{sppregnt}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab spend much more \tab spend more \tab spend same \tab spend less \tab spend much less \tab can't choose \tab no answer \tab Total \cr
   1990 \tab 155 \tab 126 \tab 400 \tab 510 \tab 84 \tab 25 \tab 53 \tab 19 \tab 1372 \cr
   Total \tab 155 \tab 126 \tab 400 \tab 510 \tab 84 \tab 25 \tab 53 \tab 19 \tab 1372 \cr
}
}

\seealso{
Other Children: 
\code{\link{chldhlth}},
\code{\link{chldhome}},
\code{\link{chldskls}},
\code{\link{chldtime}},
\code{\link{infadstv}},
\code{\link{inffilms}},
\code{\link{infmusic}},
\code{\link{infnettv}},
\code{\link{infpubtv}}
}
\concept{Children}
\keyword{variable}
