% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_05.R
\name{sppres80}
\alias{sppres80}
\title{Spouse's occupational prestige score (1980)}
\source{
General Social Survey https://gss.norc.org
}
\description{
sppres80
}
\details{
Question Spouse's Occupation, Prestige, and Industry
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010 \cr
   A/B/C/D \tab 2006 \cr
   A/Full \tab 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/32/vshow}{GSS Data Explorer page for \code{sppres80}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab not available in this year \tab 17 \tab 19 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 59 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 67 \tab 68 \tab 69 \tab 71 \tab 73 \tab 74 \tab 75 \tab 86 \tab dk, na, iap \tab 72 \tab 70 \tab no answer \tab skipped on web \tab iap \tab Total \cr
   1988 \tab - \tab 4 \tab 2 \tab 3 \tab 4 \tab 18 \tab 8 \tab 7 \tab 6 \tab 2 \tab 6 \tab 15 \tab 18 \tab 34 \tab 17 \tab 23 \tab 13 \tab 22 \tab 27 \tab 30 \tab 6 \tab 8 \tab 27 \tab 25 \tab 4 \tab 22 \tab 19 \tab 26 \tab 16 \tab 42 \tab 32 \tab 5 \tab 21 \tab 14 \tab 62 \tab 5 \tab 7 \tab 9 \tab 4 \tab 5 \tab 4 \tab 1 \tab 13 \tab 13 \tab 3 \tab 2 \tab 3 \tab 24 \tab 4 \tab 19 \tab 1 \tab 4 \tab 4 \tab 2 \tab 3 \tab 5 \tab 7 \tab 6 \tab 745 \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab - \tab - \tab 1 \tab 4 \tab 2 \tab 15 \tab 10 \tab 11 \tab 5 \tab 3 \tab 8 \tab 23 \tab 20 \tab 35 \tab 13 \tab 28 \tab 19 \tab 18 \tab 19 \tab 37 \tab 5 \tab 7 \tab 24 \tab 39 \tab 1 \tab 26 \tab 13 \tab 38 \tab 13 \tab 55 \tab 31 \tab 6 \tab 20 \tab 16 \tab 60 \tab 7 \tab 9 \tab 10 \tab 2 \tab 3 \tab 4 \tab 1 \tab 11 \tab 19 \tab 10 \tab 7 \tab 1 \tab 18 \tab 4 \tab 18 \tab 1 \tab 4 \tab 6 \tab 3 \tab - \tab 16 \tab 7 \tab 3 \tab 746 \tab 2 \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab - \tab - \tab 2 \tab 1 \tab 5 \tab 9 \tab 8 \tab 13 \tab 5 \tab - \tab 6 \tab 16 \tab 11 \tab 27 \tab 13 \tab 17 \tab 10 \tab 13 \tab 25 \tab 42 \tab 5 \tab 4 \tab 35 \tab 26 \tab 7 \tab 25 \tab 10 \tab 20 \tab 11 \tab 30 \tab 34 \tab 7 \tab 23 \tab 12 \tab 59 \tab 6 \tab 4 \tab 9 \tab 4 \tab 2 \tab 3 \tab 3 \tab 2 \tab 11 \tab 7 \tab 3 \tab 1 \tab 28 \tab 8 \tab 9 \tab - \tab 5 \tab 10 \tab - \tab 5 \tab 12 \tab 7 \tab 3 \tab 695 \tab 3 \tab 1 \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab - \tab 1 \tab - \tab 4 \tab 1 \tab 21 \tab 13 \tab 15 \tab 3 \tab 1 \tab 6 \tab 16 \tab 27 \tab 28 \tab 16 \tab 27 \tab 14 \tab 23 \tab 14 \tab 27 \tab 6 \tab 10 \tab 18 \tab 26 \tab 5 \tab 32 \tab 15 \tab 21 \tab 14 \tab 38 \tab 41 \tab 11 \tab 25 \tab 9 \tab 67 \tab 9 \tab 7 \tab 10 \tab 6 \tab - \tab 3 \tab 1 \tab 5 \tab 12 \tab 4 \tab 6 \tab 2 \tab 24 \tab 15 \tab 33 \tab - \tab 3 \tab 6 \tab - \tab 1 \tab 8 \tab 4 \tab 4 \tab 755 \tab 4 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 1 \tab 1 \tab 2 \tab 3 \tab 19 \tab 13 \tab 12 \tab 6 \tab 1 \tab 10 \tab 22 \tab 20 \tab 28 \tab 7 \tab 15 \tab 10 \tab 25 \tab 28 \tab 38 \tab 6 \tab 5 \tab 32 \tab 29 \tab 9 \tab 32 \tab 14 \tab 29 \tab 19 \tab 33 \tab 43 \tab 7 \tab 25 \tab 14 \tab 87 \tab 10 \tab 6 \tab 5 \tab 5 \tab - \tab 9 \tab 6 \tab 7 \tab 14 \tab 10 \tab 3 \tab 4 \tab 24 \tab 9 \tab 21 \tab - \tab 1 \tab 10 \tab 3 \tab 3 \tab 9 \tab 10 \tab 3 \tab 787 \tab 1 \tab 1 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab - \tab 1 \tab 3 \tab 7 \tab 7 \tab 35 \tab 14 \tab 26 \tab 7 \tab 4 \tab 12 \tab 37 \tab 44 \tab 47 \tab 34 \tab 46 \tab 29 \tab 31 \tab 41 \tab 47 \tab 9 \tab 13 \tab 34 \tab 50 \tab 22 \tab 59 \tab 35 \tab 51 \tab 29 \tab 69 \tab 65 \tab 16 \tab 61 \tab 28 \tab 116 \tab 16 \tab 10 \tab 17 \tab 1 \tab 2 \tab 12 \tab 4 \tab 20 \tab 24 \tab 18 \tab 5 \tab 4 \tab 69 \tab 17 \tab 39 \tab - \tab 3 \tab 12 \tab 5 \tab 12 \tab 23 \tab 7 \tab 4 \tab 1530 \tab 8 \tab 1 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab - \tab 7 \tab 1 \tab 2 \tab 2 \tab 33 \tab 20 \tab 26 \tab 8 \tab 2 \tab 11 \tab 28 \tab 15 \tab 56 \tab 36 \tab 52 \tab 23 \tab 26 \tab 41 \tab 64 \tab 6 \tab 12 \tab 24 \tab 48 \tab 17 \tab 46 \tab 16 \tab 44 \tab 26 \tab 55 \tab 59 \tab 13 \tab 18 \tab 18 \tab 148 \tab 15 \tab 9 \tab 12 \tab 8 \tab 2 \tab 8 \tab 4 \tab 12 \tab 25 \tab 16 \tab 4 \tab 8 \tab 60 \tab 21 \tab 39 \tab 1 \tab 9 \tab 18 \tab 5 \tab 6 \tab 15 \tab 12 \tab 9 \tab 1579 \tab 4 \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab - \tab 1 \tab 3 \tab 4 \tab 4 \tab 24 \tab 6 \tab 15 \tab 5 \tab 1 \tab 14 \tab 25 \tab 26 \tab 55 \tab 35 \tab 41 \tab 26 \tab 35 \tab 30 \tab 64 \tab 4 \tab 7 \tab 40 \tab 39 \tab 8 \tab 40 \tab 26 \tab 52 \tab 25 \tab 44 \tab 52 \tab 18 \tab 43 \tab 12 \tab 150 \tab 10 \tab 13 \tab 13 \tab 8 \tab 2 \tab 9 \tab 4 \tab 16 \tab 29 \tab 21 \tab 6 \tab 1 \tab 55 \tab 15 \tab 35 \tab 1 \tab 2 \tab 21 \tab 3 \tab 12 \tab 21 \tab 8 \tab 6 \tab 1544 \tab 2 \tab 1 \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab - \tab 5 \tab 1 \tab 3 \tab 4 \tab 19 \tab 15 \tab 20 \tab 6 \tab 5 \tab 10 \tab 19 \tab 25 \tab 69 \tab 23 \tab 29 \tab 26 \tab 28 \tab 37 \tab 39 \tab 8 \tab 12 \tab 36 \tab 51 \tab 7 \tab 41 \tab 28 \tab 44 \tab 22 \tab 40 \tab 59 \tab 7 \tab 23 \tab 18 \tab 121 \tab 17 \tab 10 \tab 9 \tab 6 \tab - \tab 11 \tab 5 \tab 7 \tab 18 \tab 6 \tab 5 \tab 7 \tab 58 \tab 23 \tab 31 \tab - \tab 2 \tab 11 \tab 1 \tab 5 \tab 25 \tab 11 \tab 7 \tab 1640 \tab 1 \tab 1 \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab - \tab 5 \tab 5 \tab 9 \tab 6 \tab 18 \tab 12 \tab 21 \tab 3 \tab 2 \tab 11 \tab 28 \tab 25 \tab 40 \tab 20 \tab 33 \tab 22 \tab 31 \tab 31 \tab 56 \tab 5 \tab 9 \tab 30 \tab 36 \tab 13 \tab 38 \tab 28 \tab 37 \tab 10 \tab 41 \tab 64 \tab 10 \tab 31 \tab 18 \tab 132 \tab 13 \tab 12 \tab 14 \tab 6 \tab 3 \tab 10 \tab 5 \tab 15 \tab 25 \tab 16 \tab 4 \tab 5 \tab 57 \tab 16 \tab 36 \tab 1 \tab 4 \tab 25 \tab 3 \tab 6 \tab 22 \tab 10 \tab 9 \tab 1566 \tab 2 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 3 \tab 1 \tab 10 \tab 1 \tab 28 \tab 9 \tab 15 \tab 7 \tab - \tab 15 \tab 14 \tab 21 \tab 56 \tab 25 \tab 39 \tab 18 \tab 24 \tab 37 \tab 66 \tab 6 \tab 6 \tab 33 \tab 29 \tab 11 \tab 49 \tab 27 \tab 59 \tab 23 \tab 41 \tab 54 \tab 22 \tab 53 \tab 24 \tab 163 \tab 18 \tab 12 \tab 13 \tab 7 \tab 5 \tab 9 \tab 3 \tab 26 \tab 23 \tab 8 \tab 6 \tab 6 \tab 75 \tab 29 \tab 40 \tab 1 \tab 5 \tab 16 \tab 3 \tab 9 \tab 28 \tab 11 \tab 13 \tab 1453 \tab 2 \tab 2 \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 7 \tab 2 \tab 9 \tab 6 \tab 40 \tab 40 \tab 30 \tab 17 \tab 2 \tab 14 \tab 24 \tab 50 \tab 83 \tab 48 \tab 56 \tab 52 \tab 49 \tab 55 \tab 79 \tab 8 \tab 16 \tab 63 \tab 59 \tab 19 \tab 63 \tab 34 \tab 69 \tab 28 \tab 52 \tab 75 \tab 22 \tab 71 \tab 27 \tab 178 \tab 24 \tab 15 \tab 39 \tab 8 \tab 5 \tab 19 \tab 4 \tab 21 \tab 34 \tab 34 \tab 10 \tab 12 \tab 81 \tab 26 \tab 93 \tab - \tab 8 \tab 24 \tab 7 \tab 10 \tab 46 \tab 7 \tab 16 \tab 2512 \tab 7 \tab 1 \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 1 \tab 1 \tab 7 \tab 2 \tab 19 \tab 8 \tab 13 \tab 4 \tab 1 \tab 5 \tab 17 \tab 26 \tab 38 \tab 22 \tab 23 \tab 22 \tab 23 \tab 26 \tab 35 \tab 2 \tab 2 \tab 20 \tab 29 \tab 10 \tab 37 \tab 24 \tab 19 \tab 16 \tab 36 \tab 39 \tab 10 \tab 27 \tab 7 \tab 97 \tab 17 \tab 5 \tab 11 \tab 5 \tab 3 \tab 6 \tab 3 \tab 12 \tab 18 \tab 14 \tab 4 \tab 10 \tab 32 \tab 9 \tab 33 \tab - \tab 7 \tab 14 \tab 3 \tab 3 \tab 19 \tab 8 \tab 2 \tab 1113 \tab 3 \tab 1 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 6 \tab 1 \tab 1 \tab 3 \tab 18 \tab 9 \tab 8 \tab 6 \tab 3 \tab 8 \tab 12 \tab 20 \tab 33 \tab 19 \tab 32 \tab 26 \tab 18 \tab 27 \tab 31 \tab 2 \tab 6 \tab 20 \tab 20 \tab 8 \tab 26 \tab 22 \tab 15 \tab 11 \tab 25 \tab 33 \tab 13 \tab 34 \tab 8 \tab 94 \tab 12 \tab 9 \tab 11 \tab 2 \tab 1 \tab 6 \tab 3 \tab 13 \tab 4 \tab 9 \tab 4 \tab 7 \tab 33 \tab 9 \tab 29 \tab 2 \tab 6 \tab 8 \tab 2 \tab 3 \tab 16 \tab 6 \tab 6 \tab 1223 \tab 2 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4024 \tab - \tab - \tab 1 \tab 7 \tab - \tab 4032 \cr
   Total \tab 0 \tab 42 \tab 24 \tab 66 \tab 50 \tab 316 \tab 185 \tab 232 \tab 88 \tab 27 \tab 136 \tab 296 \tab 348 \tab 629 \tab 328 \tab 461 \tab 310 \tab 366 \tab 438 \tab 655 \tab 78 \tab 117 \tab 436 \tab 506 \tab 141 \tab 536 \tab 311 \tab 524 \tab 263 \tab 601 \tab 681 \tab 167 \tab 475 \tab 225 \tab 1534 \tab 179 \tab 128 \tab 182 \tab 72 \tab 33 \tab 113 \tab 47 \tab 180 \tab 269 \tab 176 \tab 69 \tab 71 \tab 638 \tab 205 \tab 475 \tab 8 \tab 63 \tab 185 \tab 40 \tab 78 \tab 265 \tab 115 \tab 91 \tab 21912 \tab 41 \tab 9 \tab 1 \tab 7 \tab 0 \tab 37244 \cr
}
}

\seealso{
Other Core: 
\code{\link{cohrs2}},
\code{\link{spind}},
\code{\link{spind80}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{spoccindv}},
\code{\link{sppres}},
\code{\link{spwrkslf}}
}
\concept{Core}
\keyword{variable}
