% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_231.R
\name{sprtprsn}
\alias{sprtprsn}
\title{R consider self a spiritual person}
\source{
General Social Survey https://gss.norc.org
}
\description{
sprtprsn
}
\details{
Question To what extent do you consider yourself a spiritual person?  Are you . . .
}
\section{Values}{

\itemize{
\item \code{1} very spiritual
\item \code{2} moderately spiritual
\item \code{3} slightly spiritual
\item \code{4} not spiritual at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2121/vshow}{GSS Data Explorer page for \code{sprtprsn}}.
}

\seealso{
Other Religion and Culture: 
\code{\link{artgod}},
\code{\link{artists}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{natarts}},
\code{\link{relexp}},
\code{\link{relpersn}}
}
\concept{Religion and Culture}
\keyword{variable}
