% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_12.R
\name{spsector}
\alias{spsector}
\title{Type of college spouse attended}
\source{
General Social Survey https://gss.norc.org
}
\description{
spsector
}
\details{
Question In what year did your spouse receive their (COLLEGE NAME) degree?
I. Sector of Institution?
}
\section{Values}{

\itemize{
\item \verb{[0]} administrative unit
\item \verb{[1]} public, 4-year or above
\item \verb{[2]} private not-for-profit, 4-year or above
\item \verb{[3]} private for-profit, 4-year or above
\item \verb{[4]} public, 2-year
\item \verb{[5]} private not-for-profit, 2-year or less
\item \verb{[6]} private for-profit, 2-year
\item \verb{[7]} public, less than 2-year
\item \verb{[9]} private for-profit, less than 2-year
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/77/vshow}{GSS Data Explorer page for \code{spsector}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab public, 4-year or above \tab private not-for-profit, 4-year or above \tab private for-profit, 4-year or above \tab public, 2-year \tab administrative unit \tab private for-profit, 2-year \tab Total \cr
   2012 \tab 1706 \tab 156 \tab 79 \tab 4 \tab 29 \tab - \tab - \tab 1974 \cr
   2014 \tab 2119 \tab 243 \tab 129 \tab 10 \tab 33 \tab 2 \tab 2 \tab 2538 \cr
   2016 \tab 2459 \tab 241 \tab 124 \tab 6 \tab 37 \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   Total \tab 8632 \tab 640 \tab 332 \tab 20 \tab 99 \tab 2 \tab 2 \tab 9727 \cr
}
}

\seealso{
Other Core: 
\code{\link{barate}},
\code{\link{eftotlt}},
\code{\link{gradtounder}},
\code{\link{sector}},
\code{\link{speftotlt}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme1}},
\code{\link{voednme2}}
}
\concept{Core}
\keyword{variable}
