% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_598.R
\name{spsei10}
\alias{spsei10}
\title{R's spouse's socioeconomic index (2010)}
\source{
General Social Survey https://gss.norc.org
}
\description{
spsei10
}
\details{
Question SPSEI10 is based on the 2010 Census occupation classification. It is estimated from 539 occupational categories, using SPPRES105PLUS.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} i don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1988, 1989, 1990, 1991, 1993, 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C/D \tab 2006 \cr
   Full/- \tab 1972, 1973, 1974, 1975, 1976, 1977, 1978, 1980, 1982, 1983, 1984, 1985, 1986, 1987 \cr
}
}

\seealso{
Other Occupation: 
\code{\link{cosei10}},
\code{\link{cosei10educ}},
\code{\link{masei10}},
\code{\link{masei10educ}},
\code{\link{masei10inc}},
\code{\link{pasei10educ}},
\code{\link{pasei10inc}},
\code{\link{spsei10educ}},
\code{\link{spsei10inc}}
}
\concept{Occupation}
\keyword{variable}
