% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_159.R
\name{spsibs}
\alias{spsibs}
\title{No of siblings spouse has}
\source{
General Social Survey https://gss.norc.org
}
\description{
spsibs
}
\details{
Question How many brothers and sisters does your (husband/wife) have? (Count those born alive, but no longer living, as well as those alive now. Also include stepbrothers and stepsisters, and children adopted by his/her parents.)
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1491/vshow}{GSS Data Explorer page for \code{spsibs}}.
}

\seealso{
Other Family Mobility: 
\code{\link{kdsex1}},
\code{\link{kdsex2}},
\code{\link{kdsex3}},
\code{\link{kdsex4}},
\code{\link{kdsex5}},
\code{\link{kdsex6}},
\code{\link{kdsex7}},
\code{\link{kdsex8}},
\code{\link{sphedpre105plus}}
}
\concept{Family Mobility}
\concept{Siblings}
\keyword{variable}
