% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_485.R
\name{spwkgrwn}
\alias{spwkgrwn}
\title{Did spouse work after children left home}
\source{
General Social Survey https://gss.norc.org
}
\description{
spwkgrwn
}
\details{
Question What about your spouse/partner at that time -- did he or she work outside the home full-time, part-time, or not at all…
D. And how about after the children left home?
}
\section{Values}{

\itemize{
\item \verb{[1]} worked full-time
\item \verb{[2]} worked part-time
\item \verb{[3]} stayed home
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4426/vshow}{GSS Data Explorer page for \code{spwkgrwn}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab worked full-time \tab worked part-time \tab stayed home \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1545 \tab 456 \tab 62 \tab 69 \tab 415 \tab 445 \tab 2992 \cr
   2002 \tab 1936 \tab 344 \tab 44 \tab 36 \tab 393 \tab 12 \tab 2765 \cr
   Total \tab 3481 \tab 800 \tab 106 \tab 105 \tab 808 \tab 457 \tab 5757 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{dinner}},
\code{\link{laundry1}},
\code{\link{mehhwork}},
\code{\link{mekdcare}},
\code{\link{repairs1}},
\code{\link{shopfood}},
\code{\link{spwkbaby}},
\code{\link{spwknokd}},
\code{\link{spwrksch}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
