% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_140.R
\name{spwksup}
\alias{spwksup}
\title{Does spouse supervise anyone}
\source{
General Social Survey https://gss.norc.org
}
\description{
spwksup
}
\details{
Question In your (husband/wife/partner)'s job, does (he/she) supervise anyone who is directly responsible to (him/her)?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7093/vshow}{GSS Data Explorer page for \code{spwksup}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2018 \tab 1732 \tab 225 \tab 383 \tab 6 \tab 2 \tab - \tab 2348 \cr
   2021 \tab 3058 \tab 334 \tab 602 \tab 34 \tab - \tab 4 \tab 4032 \cr
   2022 \tab 2882 \tab 264 \tab 377 \tab 9 \tab 11 \tab 1 \tab 3544 \cr
   Total \tab 7672 \tab 823 \tab 1362 \tab 49 \tab 13 \tab 5 \tab 9924 \cr
}
}

\keyword{variable}
