% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_485.R
\name{spwrksch}
\alias{spwrksch}
\title{Did spouse work with school age children}
\source{
General Social Survey https://gss.norc.org
}
\description{
spwrksch
}
\details{
Question What about your spouse/partner at that time -- did he or she work outside the home full-time, part-time, or not at all…
C. After the youngest child started school?
}
\section{Values}{

\itemize{
\item \verb{[1]} worked full-time
\item \verb{[2]} worked part-time
\item \verb{[3]} stayed home
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4425/vshow}{GSS Data Explorer page for \code{spwrksch}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab worked full-time \tab worked part-time \tab stayed home \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1545 \tab 608 \tab 111 \tab 125 \tab 200 \tab 403 \tab 2992 \cr
   2002 \tab 1936 \tab 482 \tab 85 \tab 66 \tab 184 \tab 12 \tab 2765 \cr
   Total \tab 3481 \tab 1090 \tab 196 \tab 191 \tab 384 \tab 415 \tab 5757 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{dinner}},
\code{\link{laundry1}},
\code{\link{mehhwork}},
\code{\link{mekdcare}},
\code{\link{repairs1}},
\code{\link{shopfood}},
\code{\link{spwkbaby}},
\code{\link{spwkgrwn}},
\code{\link{spwknokd}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
