% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_05.R
\name{spwrkslf}
\alias{spwrkslf}
\title{Spouse self-emp. or works for somebody}
\source{
General Social Survey https://gss.norc.org
}
\description{
spwrkslf
}
\details{
Question A. What kind of work (does/did) your (SPOUSE) normally do? That is, what (is/was) (his/her) job called?
OCCUPATION: (See REMARKS)
B. What (does/did) (SPOUSE) actually do in that job? Tell me, what (are/were) some of (his/her) main duties?
C. What kind of place (does/did) (SPOUSE) work for?
INDUSTRY: (See REMARKS)
D. What (do/did) they (make/do)?
E. (Is/Was) (he/she) self-employed or (does/did) (he/she) work for someone else?
}
\section{Values}{

\itemize{
\item \code{1} self-employed
\item \code{2} someone else
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/-/- \tab full \cr
   1973 \tab -/-/-/- \tab full \cr
   1974 \tab -/-/-/- \tab full \cr
   1975 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1978 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1983 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/B/C/- \tab full \cr
   1989 \tab A/B/C/- \tab full \cr
   1990 \tab A/B/C/- \tab full \cr
   1991 \tab A/B/C/- \tab full \cr
   1993 \tab A/B/C/- \tab full \cr
   1994 \tab A/B/C/- \tab full \cr
   1996 \tab A/B/C/- \tab full \cr
   1998 \tab A/B/C/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
   2024 \tab A/B/C/- \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/29/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no answer \tab self-employed \tab someone else \tab don't know \tab skipped on web \tab not available in this release \tab Total \cr
   1972 \tab 580 \tab 8 \tab 120 \tab 905 \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 537 \tab 4 \tab 119 \tab 844 \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 524 \tab - \tab 109 \tab 851 \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 572 \tab 7 \tab 104 \tab 807 \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 613 \tab 5 \tab 101 \tab 780 \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 639 \tab 1 \tab 100 \tab 790 \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 654 \tab 2 \tab 118 \tab 758 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 640 \tab 3 \tab 96 \tab 729 \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 940 \tab 4 \tab 106 \tab 810 \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 705 \tab 1 \tab 128 \tab 765 \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 687 \tab 2 \tab 111 \tab 673 \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 709 \tab - \tab 126 \tab 699 \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 700 \tab 2 \tab 116 \tab 652 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 950 \tab 8 \tab 124 \tab 737 \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 742 \tab 4 \tab 103 \tab 630 \tab 2 \tab - \tab - \tab 1481 \cr
   1989 \tab 735 \tab 18 \tab 122 \tab 660 \tab 2 \tab - \tab - \tab 1537 \cr
   1990 \tab 684 \tab 9 \tab 95 \tab 582 \tab 2 \tab - \tab - \tab 1372 \cr
   1991 \tab 748 \tab 7 \tab 91 \tab 670 \tab 1 \tab - \tab - \tab 1517 \cr
   1993 \tab 781 \tab 9 \tab 122 \tab 693 \tab 1 \tab - \tab - \tab 1606 \cr
   1994 \tab 1521 \tab 6 \tab 191 \tab 1274 \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1563 \tab 20 \tab 205 \tab 1115 \tab 1 \tab - \tab - \tab 2904 \cr
   1998 \tab 1527 \tab 24 \tab 188 \tab 1092 \tab 1 \tab - \tab - \tab 2832 \cr
   2000 \tab 1608 \tab 23 \tab 153 \tab 1032 \tab 1 \tab - \tab - \tab 2817 \cr
   2002 \tab 1547 \tab 3 \tab 172 \tab 1042 \tab 1 \tab - \tab - \tab 2765 \cr
   2004 \tab 1406 \tab 12 \tab 211 \tab 1182 \tab 1 \tab - \tab - \tab 2812 \cr
   2006 \tab 2465 \tab 16 \tab 236 \tab 1788 \tab 5 \tab - \tab - \tab 4510 \cr
   2008 \tab 1077 \tab 10 \tab 118 \tab 818 \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 1187 \tab 7 \tab 97 \tab 751 \tab 2 \tab - \tab - \tab 2044 \cr
   2012 \tab 1121 \tab 2 \tab 105 \tab 743 \tab 3 \tab - \tab - \tab 1974 \cr
   2014 \tab 1434 \tab 6 \tab 130 \tab 968 \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 1692 \tab 11 \tab 149 \tab 1015 \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 1378 \tab 11 \tab 114 \tab 845 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 2111 \tab 2 \tab 230 \tab 1674 \tab - \tab 15 \tab - \tab 4032 \cr
   2022 \tab 2124 \tab 12 \tab 184 \tab 1213 \tab 4 \tab 7 \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 36901 \tab 259 \tab 4594 \tab 30587 \tab 27 \tab 22 \tab 3309 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{cohrs2}},
\code{\link{spind}},
\code{\link{spind80}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}},
\code{\link{spoccindv}},
\code{\link{sppres}},
\code{\link{sppres80}}

Other Occupation: 
\code{\link{spind}},
\code{\link{spind80}},
\code{\link{spocc}},
\code{\link{spocc10}},
\code{\link{spocc80}}
}
\concept{Core}
\concept{Job}
\concept{Occupation}
\concept{Work}
\keyword{variable}
