% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_642.R
\name{spwrkslf2}
\alias{spwrkslf2}
\title{Spouse self-employed, with employees, or work for someone else}
\source{
General Social Survey https://gss.norc.org
}
\description{
spwrkslf2
}
\details{
Question (Did/Does) your (spouse/partner) work in their own family business or farm?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, employees
\item \verb{[2]} no, no employees
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7749/vshow}{GSS Data Explorer page for \code{spwrkslf2}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes, employees \tab no, no employees \tab don't know \tab Total \cr
   2022 \tab 3360 \tab 71 \tab 112 \tab 1 \tab 3544 \cr
   Total \tab 3360 \tab 71 \tab 112 \tab 1 \tab 3544 \cr
}
}

\keyword{variable}
