% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_642.R
\name{spwrkslffam}
\alias{spwrkslffam}
\title{Spouse work at family business or on family farm}
\source{
General Social Survey https://gss.norc.org
}
\description{
spwrkslffam
}
\details{
Question Does/did your (husband/wife/spouse/partner) work in their own family business or farm?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8033/vshow}{GSS Data Explorer page for \code{spwrkslffam}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab skipped on web \tab Total \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3360 \tab 104 \tab 79 \tab 1 \tab 3544 \cr
   Total \tab 9740 \tab 104 \tab 79 \tab 1 \tab 9924 \cr
}
}

\keyword{variable}
