% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_263.R
\name{srcheng}
\alias{srcheng}
\title{How often r uses search engine}
\source{
General Social Survey https://gss.norc.org
}
\description{
srcheng
}
\details{
Question Now I’d like to ask you about how you navigate around the Web. How frequently do you use each of the following ways of getting to the web sites you visit:
C. Use a Search Engine?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} less than half of the time
\item \code{3} about half of the time
\item \code{4} more than half of the time
\item \code{5} almost always
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2390/vshow}{GSS Data Explorer page for \code{srcheng}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab never \tab less than half of the time \tab about half of the time \tab more than half of the time \tab almost always \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2075 \tab 83 \tab 231 \tab 160 \tab 94 \tab 96 \tab 1 \tab 77 \tab 2817 \cr
   2002 \tab 2403 \tab 27 \tab 125 \tab 70 \tab 42 \tab 48 \tab 1 \tab 49 \tab 2765 \cr
   Total \tab 4478 \tab 110 \tab 356 \tab 230 \tab 136 \tab 144 \tab 2 \tab 126 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{chathr}},
\code{\link{chatmin}},
\code{\link{drctlink}},
\code{\link{emailhr}},
\code{\link{emailmin}},
\code{\link{typeurl}},
\code{\link{usewww}},
\code{\link{wwwhr}},
\code{\link{wwwmin}}
}
\concept{Information Society}
\keyword{variable}
