% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_276.R
\name{srcheng2}
\alias{srcheng2}
\title{Name of search engine given by r}
\source{
General Social Survey https://gss.norc.org
}
\description{
srcheng2
}
\details{
Question Now we’d like to ask you a few questions about terms that people use to talk about the Internet and World Wide Web.
B. Please name one.
}
\section{Values}{

\itemize{
\item \code{1} yahoo
\item \code{2} lycos
\item \code{3} search.com
\item \code{4} gopher
\item \code{5} all the web
\item \code{6} msn
\item \code{7} ebay
\item \code{8} infoseek
\item \code{9} altavista
\item \code{10} explorer
\item \code{11} netscape
\item \code{12} isearch
\item \code{13} excite
\item \code{14} aol
\item \code{15} java
\item \code{16} webcrawler
\item \code{17} hotbot
\item \code{18} google
\item \code{19} copernic
\item \code{20} ask jeeves
\item \code{21} iwon
\item \code{22} dogpile
\item \code{23} streetviolence.com
\item \code{24} browser
\item \code{25} northern lights
\item \code{26} www.irs.com
\item \code{27} netfind
\item \code{28} keysearch
\item \code{29} metacrawler
\item \code{30} uproar.com
\item \code{31} siap
\item \code{32} go
\item \code{33} myway
\item \code{34} fuse
\item \code{35} espn
\item \code{36} uswest
\item \code{37} terra.com
\item \code{38} goto.com
\item \code{39} bizrate
\item \code{40} fastsearch
\item \code{41} sherlock
\item \code{42} lynco
\item \code{43} inference
\item \code{44} 4anything
\item \code{45} navistar
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2512/vshow}{GSS Data Explorer page for \code{srcheng2}}.
}

\seealso{
Other Information Society: 
\code{\link{advlib}},
\code{\link{advofrnd}},
\code{\link{cmpvirus}},
\code{\link{download}},
\code{\link{hlthinfo}},
\code{\link{hlthmag1}},
\code{\link{hlthpapr}},
\code{\link{srcheng1}},
\code{\link{upload}}
}
\concept{Information Society}
\keyword{variable}
