% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_276.R
\name{srcheng2}
\alias{srcheng2}
\title{Name of search engine given by r}
\source{
General Social Survey https://gss.norc.org
}
\description{
srcheng2
}
\details{
Question Now we'd like to ask you a few questions about terms that people use to talk about the Internet and World Wide Web.
B. Please name one.
}
\section{Values}{

\itemize{
\item \verb{[1]} yahoo
\item \verb{[2]} lycos
\item \verb{[3]} search.com
\item \verb{[4]} gopher
\item \verb{[5]} all the web
\item \verb{[6]} msn
\item \verb{[7]} ebay
\item \verb{[8]} infoseek
\item \verb{[9]} altavista
\item \verb{[10]} explorer
\item \verb{[11]} netscape
\item \verb{[12]} isearch
\item \verb{[13]} excite
\item \verb{[14]} aol
\item \verb{[15]} java
\item \verb{[16]} webcrawler
\item \verb{[17]} hotbot
\item \verb{[18]} google
\item \verb{[19]} copernic
\item \verb{[20]} ask jeeves
\item \verb{[21]} iwon
\item \verb{[22]} dogpile
\item \verb{[23]} streetviolence.com
\item \verb{[24]} browser
\item \verb{[25]} northern lights
\item \verb{[26]} www.irs.com
\item \verb{[27]} netfind
\item \verb{[28]} keysearch
\item \verb{[29]} metacrawler
\item \verb{[30]} uproar.com
\item \verb{[31]} siap
\item \verb{[32]} go
\item \verb{[33]} myway
\item \verb{[34]} fuse
\item \verb{[35]} espn
\item \verb{[36]} uswest
\item \verb{[37]} terra.com
\item \verb{[38]} goto.com
\item \verb{[39]} bizrate
\item \verb{[40]} fastsearch
\item \verb{[41]} sherlock
\item \verb{[42]} lynco
\item \verb{[43]} inference
\item \verb{[44]} 4anything
\item \verb{[45]} navistar
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2512/vshow}{GSS Data Explorer page for \code{srcheng2}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab yahoo \tab lycos \tab search.com \tab gopher \tab all the web \tab msn \tab ebay \tab infoseek \tab altavista \tab explorer \tab netscape \tab isearch \tab excite \tab aol \tab java \tab webcrawler \tab hotbot \tab google \tab copernic \tab ask jeeves \tab iwon \tab dogpile \tab streetviolence.com \tab browser \tab northern lights \tab www.irs.com \tab netfind \tab keysearch \tab metacrawler \tab uproar.com \tab siap \tab go \tab myway \tab fuse \tab espn \tab goto.com \tab bizrate \tab sherlock \tab lynco \tab 4anything \tab navistar \tab don't know \tab no answer \tab Total \cr
   2000 \tab 2215 \tab 315 \tab 33 \tab 1 \tab 1 \tab 1 \tab 7 \tab 1 \tab 10 \tab 26 \tab 7 \tab 14 \tab 1 \tab 19 \tab 17 \tab 1 \tab 4 \tab 7 \tab 12 \tab 1 \tab 3 \tab 5 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 82 \tab 2817 \cr
   Total \tab 2215 \tab 315 \tab 33 \tab 1 \tab 1 \tab 1 \tab 7 \tab 1 \tab 10 \tab 26 \tab 7 \tab 14 \tab 1 \tab 19 \tab 17 \tab 1 \tab 4 \tab 7 \tab 12 \tab 1 \tab 3 \tab 5 \tab 6 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 2 \tab 1 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 82 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{advlib}},
\code{\link{advofrnd}},
\code{\link{cmpvirus}},
\code{\link{download}},
\code{\link{hlthinfo}},
\code{\link{hlthmag1}},
\code{\link{hlthpapr}},
\code{\link{srcheng1}},
\code{\link{upload}}
}
\concept{Information Society}
\keyword{variable}
