% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_146.R
\name{standup}
\alias{standup}
\title{Stand up for your own opinion}
\source{
General Social Survey https://gss.norc.org
}
\description{
standup
}
\details{
Question Some people think that it's important to stand up for your own opinion even if it makes others around you  uncomfortable. Others think that it's better to keep your views to yourself if they would make others around you uncomfortable. Which position comes closer to your view?
}
\section{Values}{

\itemize{
\item \code{1} should stand up regardless
\item \code{2} keep it yourself if offend others
\item \code{3} neither
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1376/vshow}{GSS Data Explorer page for \code{standup}}.
}

\seealso{
Other Culture: 
\code{\link{richpoor}},
\code{\link{selfirst}}
}
\concept{Culture}
\concept{Sociability}
\keyword{variable}
