% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_417.R
\name{startbiz}
\alias{startbiz}
\title{R trying to start business}
\source{
General Social Survey https://gss.norc.org
}
\description{
startbiz
}
\details{
Question Are you, alone or with others, currently trying to start a new business, including any self-employment, or selling any goods or services to others?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3814/vshow}{GSS Data Explorer page for \code{startbiz}}.
}

\seealso{
Other Self Employment: 
\code{\link{bizgross}},
\code{\link{bizshare}},
\code{\link{numemps5}},
\code{\link{numown}},
\code{\link{ownbiz}},
\code{\link{owninc}},
\code{\link{paidemps}},
\code{\link{whybiz}},
\code{\link{yearbiz}}
}
\concept{Self Employment}
\keyword{variable}
