% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_232.R
\name{stateart}
\alias{stateart}
\title{Feel state gov should provide assistance to art}
\source{
General Social Survey https://gss.norc.org
}
\description{
stateart
}
\details{
Question If arts organizations -- such as art museums, dance, operas, theater groups and symphony orchestras --  need financial assistance to operate, Do you feel that the federal government should provide assistance or not?
A. What about state government? Do you feel that it should provide financial assistance to arts organizations?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2135/vshow}{GSS Data Explorer page for \code{stateart}}.
}

\seealso{
Other Religion and Culture: 
\code{\link{aimofart}},
\code{\link{artgod}},
\code{\link{artists}},
\code{\link{beausprt}},
\code{\link{conghlp1}},
\code{\link{conghlp2}},
\code{\link{conghrm1}},
\code{\link{conghrm2}},
\code{\link{cope1}},
\code{\link{cope2}},
\code{\link{cope3}},
\code{\link{cope4}},
\code{\link{cope5}},
\code{\link{cope6}},
\code{\link{drama}},
\code{\link{effort}},
\code{\link{feelgod}},
\code{\link{fidgety}},
\code{\link{forgive1}},
\code{\link{forgive2}},
\code{\link{forgive3}},
\code{\link{godlove}},
\code{\link{godwatch}},
\code{\link{harmony}},
\code{\link{hopeless}},
\code{\link{irrelart}},
\code{\link{irreloff}},
\code{\link{lesspain}},
\code{\link{localart}},
\code{\link{meditate}},
\code{\link{natarts}},
\code{\link{natlart}},
\code{\link{nervous}},
\code{\link{nocheer}},
\code{\link{popmusic}},
\code{\link{privpray}},
\code{\link{readfict}},
\code{\link{relart}},
\code{\link{relcmfrt}},
\code{\link{relexp}},
\code{\link{rellife}},
\code{\link{relpersn}},
\code{\link{sprtprsn}},
\code{\link{uniongod}},
\code{\link{volarts}},
\code{\link{wrthless}}
}
\concept{Religion and Culture}
\keyword{variable}
