% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_504.R
\name{stayorg3}
\alias{stayorg3}
\title{I would turn down job offering more pay}
\source{
General Social Survey https://gss.norc.org
}
\description{
stayorg3
}
\details{
Question To what extent do you agree or disagree with each of the following statements?
D. I would turn down another job that offered quite a bit more pay in order to stay with this organization.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4588/vshow}{GSS Data Explorer page for \code{stayorg3}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1967 \tab 45 \tab 122 \tab 201 \tab 288 \tab 136 \tab 29 \tab 44 \tab 2832 \cr
   2006 \tab 3492 \tab 128 \tab 201 \tab 181 \tab 314 \tab 187 \tab 5 \tab 2 \tab 4510 \cr
   2016 \tab 1927 \tab 108 \tab 146 \tab 190 \tab 282 \tab 201 \tab 5 \tab 8 \tab 2867 \cr
   Total \tab 7386 \tab 281 \tab 469 \tab 572 \tab 884 \tab 524 \tab 39 \tab 54 \tab 10209 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{absent}},
\code{\link{chngwork}},
\code{\link{edcskill}},
\code{\link{helporg1}},
\code{\link{jobskill}},
\code{\link{prideorg}},
\code{\link{proudwrk}},
\code{\link{useskill}},
\code{\link{wktenure}}
}
\concept{Work Orientation}
\keyword{variable}
