% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_651.R
\name{stockyr}
\alias{stockyr}
\title{Granted shares of stock in past year}
\source{
General Social Survey https://gss.norc.org
}
\description{
stockyr
}
\details{
Question Were you granted shares of stock in the company where you work in 2021? This means shares that the company gave to you for free through a restricted stock plan or an Employee Stock Ownership Plan, a stock bonus plan, or other stock plan. It does not mea
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7846/vshow}{GSS Data Explorer page for \code{stockyr}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab skipped on web \tab Total \cr
   2022 \tab 3423 \tab 46 \tab 67 \tab 6 \tab 2 \tab 3544 \cr
   Total \tab 3423 \tab 46 \tab 67 \tab 6 \tab 2 \tab 3544 \cr
}
}

\keyword{variable}
