% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_651.R
\name{stockyrval}
\alias{stockyrval}
\title{Value of stock gratned in past year}
\source{
General Social Survey https://gss.norc.org
}
\description{
stockyrval
}
\details{
Question Please give a general estimate of the total value of all shares of company stock that your company granted to you in 2021 when they were granted to you.
}
\section{Values}{

\itemize{
\item \code{1} less than $499
\item \code{2} $500 to $999
\item \code{3} $1,000 to $2,999
\item \code{4} $3,000 to $4,999
\item \code{5} $5,000 to $9,999
\item \code{6} $10,000 to $19,999
\item \code{7} $20,000 to $39,999
\item \code{8} $40,000 to $79,999
\item \code{9} $80,000 to $199,999
\item \code{10} $200,000 to $499,999
\item \code{11} $500,000 to $999,000
\item \code{12} over $1 million
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7847/vshow}{GSS Data Explorer page for \code{stockyrval}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab less than $499 \tab $500 to $999 \tab $1,000 to $2,999 \tab $3,000 to $4,999 \tab $5,000 to $9,999 \tab $10,000 to $19,999 \tab $20,000 to $39,999 \tab $40,000 to $79,999 \tab $80,000 to $199,999 \tab over $1 million \tab don't know \tab no answer \tab Total \cr
   2022 \tab 3498 \tab 5 \tab 4 \tab 8 \tab 4 \tab 3 \tab 3 \tab 2 \tab 5 \tab 2 \tab 1 \tab 8 \tab 1 \tab 3544 \cr
   Total \tab 3498 \tab 5 \tab 4 \tab 8 \tab 4 \tab 3 \tab 3 \tab 2 \tab 5 \tab 2 \tab 1 \tab 8 \tab 1 \tab 3544 \cr
}
}

\keyword{variable}
