% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_652.R
\name{stoptyramt}
\alias{stoptyramt}
\title{Value of options granted in past year}
\source{
General Social Survey https://gss.norc.org
}
\description{
stoptyramt
}
\details{
Question Please give a general estimate of the total number of the stock options that your company granted to you in 2021.
}
\section{Values}{

\itemize{
\item \verb{[1]} less than 100 stock options
\item \verb{[2]} 100 to 499 stock options
\item \verb{[3]} 500 to 999 stock options
\item \verb{[4]} 1,000 to 2,999 stock options
\item \verb{[5]} 3,000 to 4,999 stock options
\item \verb{[6]} 5,000 to 9,999 stock options
\item \verb{[7]} 10,000 to 19,999 stock options
\item \verb{[8]} 20,000 to 39,999 stock options
\item \verb{[9]} 40,000 to 79,999 stock options
\item \verb{[10]} 80,000 to 99,999 stock options
\item \verb{[11]} 100,000 to 199,999 stock options
\item \verb{[12]} 200,000 to 299,000 stock options
\item \verb{[13]} 300,000 to 399,999 stock options
\item \verb{[14]} 400,000 to 499,999 stock options
\item \verb{[15]} 500,000 to 699,999 stock options
\item \verb{[16]} 700,000 to 999,999 stock options
\item \verb{[17]} over one million stock options
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7849/vshow}{GSS Data Explorer page for \code{stoptyramt}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab less than 100 stock options \tab 100 to 499 stock options \tab 500 to 999 stock options \tab 1,000 to 2,999 stock options \tab 5,000 to 9,999 stock options \tab 40,000 to 79,999 stock options \tab don't know \tab Total \cr
   2022 \tab 3529 \tab 6 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3544 \cr
   Total \tab 3529 \tab 6 \tab 2 \tab 1 \tab 1 \tab 3 \tab 1 \tab 1 \tab 3544 \cr
}
}

\keyword{variable}
