% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_383.R
\name{stormtxt}
\alias{stormtxt}
\title{Do you see lightning before hearing thunder}
\source{
General Social Survey https://gss.norc.org
}
\description{
stormtxt
}
\details{
Question Lightning and thunder happen at the same time, but you see the lightning before you hear the thunder. Explain why this is so.
}
\section{Values}{

\itemize{
\item \verb{[1]} complete
\item \verb{[2]} partial
\item \verb{[3]} unsatisfactory/incorrect
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3474/vshow}{GSS Data Explorer page for \code{stormtxt}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab complete \tab partial \tab unsatisfactory/incorrect \tab don't know \tab no answer \tab Total \cr
   2008 \tab 518 \tab 656 \tab 61 \tab 299 \tab 466 \tab 23 \tab 2023 \cr
   Total \tab 518 \tab 656 \tab 61 \tab 299 \tab 466 \tab 23 \tab 2023 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{anheat}},
\code{\link{anscitst}},
\code{\link{erosion}},
\code{\link{goldfish}},
\code{\link{h2olife}},
\code{\link{lftplane}},
\code{\link{litmstxt}},
\code{\link{maboygrl}},
\code{\link{salth2o}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
