% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_312.R
\name{stress12}
\alias{stress12}
\title{Stress management program last 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
stress12
}
\details{
Question In the past 12 months, have you participated in any of these stress management or stress reduction programs?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2841/vshow}{GSS Data Explorer page for \code{stress12}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   2014 \tab 1288 \tab 81 \tab 397 \tab 772 \tab - \tab - \tab 2538 \cr
   2018 \tab 900 \tab 81 \tab 429 \tab 938 \tab - \tab - \tab 2348 \cr
   2022 \tab 2575 \tab 183 \tab 694 \tab 1 \tab 68 \tab 23 \tab 3544 \cr
   Total \tab 4763 \tab 345 \tab 1520 \tab 1711 \tab 68 \tab 23 \tab 8430 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{arthrtis}},
\code{\link{depress}},
\code{\link{diabetes}},
\code{\link{hyperten}},
\code{\link{knowschd}},
\code{\link{satjob1}},
\code{\link{slpprblm}},
\code{\link{usetech}},
\code{\link{weight}}
}
\concept{Quality of Working Life}
\keyword{variable}
