% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_210.R
\name{stresses}
\alias{stresses}
\title{How likely x s situation caused by stress}
\source{
General Social Survey https://gss.norc.org
}
\description{
stresses
}
\details{
Question In your opinion, how likely is it that NAME's situation might be caused by (READ a-f) -- very likely, somewhat likely, not very likely, or not at all likely?
D. Stressful circumstances in his/her life
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2018 \tab B/C/- \tab partial \cr
   2024 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1991/vshow}{GSS Data Explorer page for \code{stresses}}.
}

\seealso{
Other Mental Health: 
\code{\link{charactr}},
\code{\link{imbalnce}},
\code{\link{probfix3}},
\code{\link{probhlp1}},
\code{\link{probhlp2}},
\code{\link{probhlp3}},
\code{\link{seriousp}},
\code{\link{vigversn}},
\code{\link{wayraise}}
}
\concept{Mental Health}
\keyword{variable}
