% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_86.R
\name{stresswk}
\alias{stresswk}
\title{Stress in past 2 wks}
\source{
General Social Survey https://gss.norc.org
}
\description{
stresswk
}
\details{
Question During the past two weeks, would you say that you experienced a lot of stress, a moderate amount of stress, relatively little stress, or almost no stress at all?
}
\section{Values}{

\itemize{
\item \code{1} a lot of stress
\item \code{2} a moderate amount of stress
\item \code{3} relatively little stress
\item \code{4} almost no stress at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6196/vshow}{GSS Data Explorer page for \code{stresswk}}.
}

\seealso{
Other Core: 
\code{\link{bored}},
\code{\link{civic}},
\code{\link{feserve}},
\code{\link{imppol}},
\code{\link{meserve}},
\code{\link{pollgood}},
\code{\link{polltrue}},
\code{\link{privacy}},
\code{\link{rushed}}
}
\concept{Core}
\keyword{variable}
