% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_493.R
\name{strngun}
\alias{strngun}
\title{Strong unions needed to protect workers}
\source{
General Social Survey https://gss.norc.org
}
\description{
strngun
}
\details{
Question How much do you agree or disagree with these two statements?
B. Workers need strong trade unions to protect their interests.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
   2016 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4506/vshow}{GSS Data Explorer page for \code{strngun}}.
}

\seealso{
Other Work Orientation: 
\code{\link{bosswrks}},
\code{\link{dk}},
\code{\link{educemp}},
\code{\link{famresp}},
\code{\link{hiinc}},
\code{\link{intjob}},
\code{\link{leisure}},
\code{\link{promotn}},
\code{\link{secjob}}
}
\concept{Labor Unions}
\concept{Work Orientation}
\keyword{variable}
