% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_636.R
\name{strvbiasy}
\alias{strvbiasy}
\title{Bigger priority for criminal justice system}
\source{
General Social Survey https://gss.norc.org
}
\description{
strvbiasy
}
\details{
Question Which do you think should be the bigger priority for the U.S. criminal justice system today?
}
\section{Values}{

\itemize{
\item \code{1} strengthening law and order through more police and greater enforcement of the laws
\item \code{2} reducing bias against minorities in the criminal justice system by reforming court and police practices
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7576/vshow}{GSS Data Explorer page for \code{strvbiasy}}.
}

\keyword{variable}
