% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_55.R
\name{studious}
\alias{studious}
\title{Good student}
\source{
General Social Survey https://gss.norc.org
}
\description{
studious
}
\details{
Question A. Which three qualities listed on this card would you say are the most desirable for a child to have?
B. Which one of these three is the most desirable of all?
C. All of the qualities listed on this card may be desirable, but could you tell me which three you consider least important?
D. And which one of these three is least important of all?
13. That he is a good student
}
\section{Values}{

\itemize{
\item \code{1} 1 most desirable
\item \code{2} 3 most desirable
\item \code{3} not mentioned
\item \code{4} 3 least desirable
\item \code{5} 1 least desirable
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab - \tab full \cr
   1975 \tab - \tab full \cr
   1976 \tab - \tab full \cr
   1978 \tab - \tab full \cr
   1980 \tab - \tab full \cr
   1983 \tab - \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/486/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab 1 least desirable \tab 1 most desirable \tab 3 least desirable \tab 3 most desirable \tab no answer \tab not mentioned \tab don't know \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 157 \tab 15 \tab 444 \tab 62 \tab 4 \tab 822 \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 130 \tab 13 \tab 462 \tab 55 \tab 9 \tab 821 \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 120 \tab 6 \tab 447 \tab 67 \tab 9 \tab 850 \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 120 \tab 7 \tab 496 \tab 67 \tab 10 \tab 830 \tab 2 \tab - \tab 1532 \cr
   1980 \tab 975 \tab 32 \tab 2 \tab 147 \tab 25 \tab 3 \tab 283 \tab 1 \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 120 \tab 13 \tab 473 \tab 82 \tab 20 \tab 891 \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 60729 \tab 679 \tab 56 \tab 2469 \tab 358 \tab 55 \tab 4497 \tab 3 \tab 6853 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{amicably}},
\code{\link{cleany}},
\code{\link{controly}},
\code{\link{honesty}},
\code{\link{interest}},
\code{\link{judgmeny}},
\code{\link{mannersy}},
\code{\link{roley}},
\code{\link{successy}}

Other Children: 
\code{\link{amicably}},
\code{\link{cleany}},
\code{\link{controly}},
\code{\link{honesty}},
\code{\link{interest}},
\code{\link{judgmeny}},
\code{\link{mannersy}},
\code{\link{roley}},
\code{\link{successy}}

Other Split Ballots: 
\code{\link{amicably}},
\code{\link{cleany}},
\code{\link{controly}},
\code{\link{honesty}},
\code{\link{interest}},
\code{\link{judgmeny}},
\code{\link{mannersy}},
\code{\link{roley}},
\code{\link{successy}}
}
\concept{Children}
\concept{Core}
\concept{Split Ballots}
\keyword{variable}
