% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_669.R
\name{subsamprate}
\alias{subsamprate}
\title{Subsampling rate corresponding to two-phase design}
\source{
General Social Survey https://gss.norc.org
}
\description{
subsamprate
}
\details{
Question Subsampling rate corresponding to two-phase design
}
\section{Values}{

\itemize{
\item \code{1} 100pct subsampling rate
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
   2024 \tab A/B/C/- \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8042/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllll}{
   year \tab not available in this year \tab 0.5 \tab iap \tab 0.2 \tab 100pct subsampling rate \tab 0.3 \tab 0.4 \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 448 \tab 2364 \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab - \tab 874 \tab 3636 \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab - \tab 411 \tab 1612 \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab - \tab 386 \tab 1658 \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab - \tab 488 \tab 1486 \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab - \tab 807 \tab 1731 \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab - \tab 573 \tab 2294 \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab - \tab 248 \tab 2100 \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab 1326 \tab 744 \tab 1474 \tab - \tab - \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab 532 \tab 1511 \tab 577 \tab 689 \tab 3309 \cr
   Total \tab 47730 \tab 4235 \tab 18207 \tab 1276 \tab 2985 \tab 577 \tab 689 \tab 75699 \cr
}
}

\seealso{
Other Sampling and Weights: 
\code{\link{batch}}
}
\concept{Sampling and Weights}
\keyword{variable}
