% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_550.R
\name{suiage2}
\alias{suiage2}
\title{Age of suicide victim 2nd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
suiage2
}
\details{
Question We would like to know a few other things about that person.
B. How old was that person?  Was (he/she)…
}
\section{Values}{

\itemize{
\item \verb{[1]} 10 years or under
\item \verb{[2]} 11-20
\item \verb{[3]} 21-40
\item \verb{[4]} 41 year or older
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5033/vshow}{GSS Data Explorer page for \code{suiage2}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab 11-20 \tab 21-40 \tab 41 year or older \tab don't know \tab no answer \tab Total \cr
   1990 \tab 1347 \tab 6 \tab 11 \tab 5 \tab 1 \tab 2 \tab 1372 \cr
   1991 \tab 1477 \tab 7 \tab 16 \tab 14 \tab - \tab 3 \tab 1517 \cr
   Total \tab 2824 \tab 13 \tab 27 \tab 19 \tab 1 \tab 5 \tab 2889 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{suiage}},
\code{\link{suirac2}},
\code{\link{suirace}},
\code{\link{suireg}},
\code{\link{suireg2}},
\code{\link{suisex2}},
\code{\link{suisex3}},
\code{\link{suiwho2}},
\code{\link{suiwho3}}
}
\concept{Victims & AIDS}
\keyword{variable}
