% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_550.R
\name{suiage2}
\alias{suiage2}
\title{Age of suicide victim 2nd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
suiage2
}
\details{
Question We would like to know a few other things about that person.
B. How old was that person?  Was (he/she)…
}
\section{Values}{

\itemize{
\item \code{1} 10 years or under
\item \code{2} 11-20
\item \code{3} 21-40
\item \code{4} 41 year or older
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5033/vshow}{GSS Data Explorer page for \code{suiage2}}.
}

\seealso{
Other Victims & AIDS: 
\code{\link{suiage}},
\code{\link{suirac2}},
\code{\link{suirace}},
\code{\link{suireg}},
\code{\link{suireg2}},
\code{\link{suisex2}},
\code{\link{suisex3}},
\code{\link{suiwho2}},
\code{\link{suiwho3}}

Other Suicide: 
\code{\link{suiage}},
\code{\link{suirac2}},
\code{\link{suirace}},
\code{\link{suireg}},
\code{\link{suireg2}},
\code{\link{suisex2}},
\code{\link{suisex3}},
\code{\link{suiwho2}},
\code{\link{suiwho3}}
}
\concept{Suicide}
\concept{Victims & AIDS}
\keyword{variable}
