% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_627.R
\name{suicide2g}
\alias{suicide2g}
\title{Suicide if bankrupt}
\source{
General Social Survey https://gss.norc.org
}
\description{
suicide2g
}
\details{
Question Do you think a person has the right to end his or her own life if this person...

has gone bankrupt?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2021, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7473/vshow}{GSS Data Explorer page for \code{suicide2g}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab skipped on web \tab don't know \tab no answer \tab Total \cr
   2021 \tab 2794 \tab 133 \tab 1059 \tab 46 \tab - \tab - \tab 4032 \cr
   2022 \tab 2339 \tab 169 \tab 1001 \tab 10 \tab 18 \tab 7 \tab 3544 \cr
   Total \tab 5133 \tab 302 \tab 2060 \tab 56 \tab 18 \tab 7 \tab 7576 \cr
}
}

\keyword{variable}
