% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_73.R
\name{suiclse1}
\alias{suiclse1}
\title{How close were you to closest suicide victim}
\source{
General Social Survey https://gss.norc.org
}
\description{
suiclse1
}
\details{
Question On a scale of 1 to 5 how close did you feel to that person at the time of his or her death? Use 1 for not close at all and 5 for very close.              
(HAND CARD I3)
}
\section{Values}{

\itemize{
\item \code{1} 1 - not close at all
\item \code{5} 5 - very close
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6185/vshow}{GSS Data Explorer page for \code{suiclse1}}.
}

\seealso{
Other Suicide: 
\code{\link{suiclse2}},
\code{\link{suimths1}},
\code{\link{suimths2}},
\code{\link{suineg2}},
\code{\link{suirel2}},
\code{\link{suitime1}},
\code{\link{suitime2}},
\code{\link{suiyrs1}},
\code{\link{suiyrs2}}
}
\concept{Suicide}
\keyword{variable}
