% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_73.R
\name{suimths1}
\alias{suimths1}
\title{How many months since closest suicide}
\source{
General Social Survey https://gss.norc.org
}
\description{
suimths1
}
\details{
Question MONTHS:
}
\section{Values}{

\itemize{
\item \code{0} less than one
\item \code{97} no month given
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6187/vshow}{GSS Data Explorer page for \code{suimths1}}.

Counts by year:\tabular{llllllllllllllllll}{
   year \tab iap \tab less than one \tab 1 \tab 10 \tab 11 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab no month given \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2093 \tab 221 \tab 14 \tab 4 \tab 9 \tab 19 \tab 6 \tab 11 \tab 11 \tab 30 \tab 5 \tab 3 \tab 5 \tab 395 \tab 8 \tab 33 \tab 2867 \cr
   Total \tab 2093 \tab 221 \tab 14 \tab 4 \tab 9 \tab 19 \tab 6 \tab 11 \tab 11 \tab 30 \tab 5 \tab 3 \tab 5 \tab 395 \tab 8 \tab 33 \tab 2867 \cr
}
}

\seealso{
Other Suicide: 
\code{\link{suiclse1}},
\code{\link{suiclse2}},
\code{\link{suimths2}},
\code{\link{suineg2}},
\code{\link{suirel2}},
\code{\link{suitime1}},
\code{\link{suitime2}},
\code{\link{suiyrs1}},
\code{\link{suiyrs2}}
}
\concept{Suicide}
\keyword{variable}
