% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_72.R
\name{suineg1}
\alias{suineg1}
\title{Did closest suicide negatively impact you}
\source{
General Social Survey https://gss.norc.org
}
\description{
suineg1
}
\details{
Question Was that person's death emotionally distressing to you?
(HAND CARD I2)
}
\section{Values}{

\itemize{
\item \code{1} yes, greatly
\item \code{2} yes, to some extent
\item \code{3} yes, but not much
\item \code{4} no
\item \code{5} not sure
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6184/vshow}{GSS Data Explorer page for \code{suineg1}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab yes, greatly \tab yes, to some extent \tab yes, but not much \tab no \tab not sure \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2093 \tab 259 \tab 257 \tab 113 \tab 103 \tab 18 \tab 1 \tab 23 \tab 2867 \cr
   Total \tab 2093 \tab 259 \tab 257 \tab 113 \tab 103 \tab 18 \tab 1 \tab 23 \tab 2867 \cr
}
}

\seealso{
Other Suicide: 
\code{\link{suinum}},
\code{\link{suinum1}},
\code{\link{suirel1}}
}
\concept{Suicide}
\keyword{variable}
