% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_73.R
\name{suineg2}
\alias{suineg2}
\title{Did 2nd closest suicide negatively impact you}
\source{
General Social Survey https://gss.norc.org
}
\description{
suineg2
}
\details{
Question Was that person's death emotionally distressing to you?
(HAND CARD I5)
}
\section{Values}{

\itemize{
\item \verb{[1]} yes, greatly
\item \verb{[2]} yes, to some extent
\item \verb{[3]} yes, but not much
\item \verb{[4]} no
\item \verb{[5]} not sure
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6190/vshow}{GSS Data Explorer page for \code{suineg2}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes, greatly \tab yes, to some extent \tab yes, but not much \tab no \tab not sure \tab no answer \tab Total \cr
   2016 \tab 2425 \tab 85 \tab 138 \tab 102 \tab 84 \tab 9 \tab 24 \tab 2867 \cr
   Total \tab 2425 \tab 85 \tab 138 \tab 102 \tab 84 \tab 9 \tab 24 \tab 2867 \cr
}
}

\seealso{
Other Suicide: 
\code{\link{suiclse1}},
\code{\link{suiclse2}},
\code{\link{suimths1}},
\code{\link{suimths2}},
\code{\link{suirel2}},
\code{\link{suitime1}},
\code{\link{suitime2}},
\code{\link{suiyrs1}},
\code{\link{suiyrs2}}
}
\concept{Suicide}
\keyword{variable}
