% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_72.R
\name{suinum}
\alias{suinum}
\title{People r knows suicide over lifetime}
\source{
General Social Survey https://gss.norc.org
}
\description{
suinum
}
\details{
Question The next questions deal with a difficult topic that may be sensitive for you. It is not clearly understood how a relative or friend's death by suicide may affect one's life afterwards. Please answer these questions as best as you can, and you can decline to answer any question if you are uncomfortable.            
Over your lifetime how many people have you known personally that died by suicide? Please tell us the total number of people that died by suicide.            
PROBE:
IF R REPORTS NUMBER OF KNOWN SUICIDE VICTIMS=0, ASK: Just to be sure, thinking about anyone you may have known in your lifetime, is there anyone you have known personally that died by suicide?
NUMBER OF PEOPLE:
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6181/vshow}{GSS Data Explorer page for \code{suinum}}.

Counts by year:\tabular{lllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 12 \tab 2 \tab 20 \tab 21 \tab 3 \tab 4 \tab 5 \tab 50 \tab 6 \tab 7 \tab 70 \tab 8 \tab 95 \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1409 \tab 681 \tab 332 \tab 9 \tab 1 \tab 215 \tab 1 \tab 1 \tab 100 \tab 36 \tab 25 \tab 1 \tab 15 \tab 7 \tab 1 \tab 4 \tab 3 \tab 3 \tab 23 \tab 2867 \cr
   Total \tab 1409 \tab 681 \tab 332 \tab 9 \tab 1 \tab 215 \tab 1 \tab 1 \tab 100 \tab 36 \tab 25 \tab 1 \tab 15 \tab 7 \tab 1 \tab 4 \tab 3 \tab 3 \tab 23 \tab 2867 \cr
}
}

\seealso{
Other Suicide: 
\code{\link{suineg1}},
\code{\link{suinum1}},
\code{\link{suirel1}}
}
\concept{Suicide}
\keyword{variable}
