% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_72.R
\name{suinum}
\alias{suinum}
\title{People r knows suicide over lifetime}
\source{
General Social Survey https://gss.norc.org
}
\description{
suinum
}
\details{
Question The next questions deal with a difficult topic that may be sensitive for you. It is not clearly understood how a relative or friend's death by suicide may affect one's life afterwards. Please answer these questions as best as you can, and you can decline to answer any question if you are uncomfortable.            
Over your lifetime how many people have you known personally that died by suicide? Please tell us the total number of people that died by suicide.            
PROBE:
IF R REPORTS NUMBER OF KNOWN SUICIDE VICTIMS=0, ASK: Just to be sure, thinking about anyone you may have known in your lifetime, is there anyone you have known personally that died by suicide?
NUMBER OF PEOPLE:
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6181/vshow}{GSS Data Explorer page for \code{suinum}}.
}

\seealso{
Other Suicide: 
\code{\link{suiclse1}},
\code{\link{suiclse2}},
\code{\link{suimths1}},
\code{\link{suimths2}},
\code{\link{suineg1}},
\code{\link{suineg2}},
\code{\link{suinum1}},
\code{\link{suirel1}},
\code{\link{suirel2}},
\code{\link{suitime1}},
\code{\link{suitime2}},
\code{\link{suiyrs1}},
\code{\link{suiyrs2}}
}
\concept{Suicide}
\keyword{variable}
