% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_72.R
\name{suinum1}
\alias{suinum1}
\title{People r knows suicide over lifetime}
\source{
General Social Survey https://gss.norc.org
}
\description{
suinum1
}
\details{
Question Have you known personally two or more people, one person, or no people that that died by suicide over your lifetime?
PROBE:
IF R REPORTS NUMBER OF KNOWN SUICIDE VICTIMS=0, ASK: Just to be sure, thinking about anyone you may have known in your lifetime, is there anyone you have known personally that died by suicide?
}
\section{Values}{

\itemize{
\item \verb{[1]} one person
\item \verb{[2]} two or more people
\item \verb{[3]} no people
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6182/vshow}{GSS Data Explorer page for \code{suinum1}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab one person \tab no people \tab don't know \tab Total \cr
   2016 \tab 2858 \tab 3 \tab 3 \tab 3 \tab 2867 \cr
   Total \tab 2858 \tab 3 \tab 3 \tab 3 \tab 2867 \cr
}
}

\seealso{
Other Suicide: 
\code{\link{suineg1}},
\code{\link{suinum}},
\code{\link{suirel1}}
}
\concept{Suicide}
\keyword{variable}
