% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_550.R
\name{suirac2}
\alias{suirac2}
\title{Race of suicide victim 2nd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
suirac2
}
\details{
Question We would like to know a few other things about that person.
C. What was that person''s race?  Was it black, white, hispanic, or other?
}
\section{Values}{

\itemize{
\item \code{1} black
\item \code{2} white
\item \code{3} hispanic
\item \code{4} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5034/vshow}{GSS Data Explorer page for \code{suirac2}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab black \tab white \tab don't know \tab no answer \tab hispanic \tab other \tab Total \cr
   1990 \tab 1347 \tab 3 \tab 19 \tab 1 \tab 2 \tab - \tab - \tab 1372 \cr
   1991 \tab 1477 \tab 3 \tab 31 \tab - \tab 3 \tab 2 \tab 1 \tab 1517 \cr
   Total \tab 2824 \tab 6 \tab 50 \tab 1 \tab 5 \tab 2 \tab 1 \tab 2889 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{suiage}},
\code{\link{suiage2}},
\code{\link{suirace}},
\code{\link{suireg}},
\code{\link{suireg2}},
\code{\link{suisex2}},
\code{\link{suisex3}},
\code{\link{suiwho2}},
\code{\link{suiwho3}}
}
\concept{Victims & AIDS}
\keyword{variable}
