% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_551.R
\name{suirac3}
\alias{suirac3}
\title{Race of suicide victim 3rd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
suirac3
}
\details{
Question We would like to know a few other things about that person.
C. What was that person''s race?  Was it black, white, hispanic, or other?
}
\section{Values}{

\itemize{
\item \verb{[1]} black
\item \verb{[2]} white
\item \verb{[3]} hispanic
\item \verb{[4]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5039/vshow}{GSS Data Explorer page for \code{suirac3}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab white \tab no answer \tab hispanic \tab Total \cr
   1990 \tab 1364 \tab 6 \tab 2 \tab - \tab 1372 \cr
   1991 \tab 1502 \tab 10 \tab 3 \tab 2 \tab 1517 \cr
   Total \tab 2866 \tab 16 \tab 5 \tab 2 \tab 2889 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsads}},
\code{\link{aidshlth}},
\code{\link{aidsids}},
\code{\link{aidsinsr}},
\code{\link{aidsmar}},
\code{\link{aidssch}},
\code{\link{aidssxed}},
\code{\link{suiage3}},
\code{\link{suireg3}}
}
\concept{Victims & AIDS}
\keyword{variable}
