% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_550.R
\name{suireg2}
\alias{suireg2}
\title{Region of suicide victim 2nd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
suireg2
}
\details{
Question We would like to know a few other things about that person.
D. What state did this person live in?
}
\section{Values}{

\itemize{
\item \code{0} foreign
\item \code{1} new england
\item \code{2} middle atlantic
\item \code{3} e. nor. central
\item \code{4} w. nor. central
\item \code{5} south atlantic
\item \code{6} e. sou. central
\item \code{7} w. sou. central
\item \code{8} mountain
\item \code{9} pacific
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5035/vshow}{GSS Data Explorer page for \code{suireg2}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab new england \tab e. nor. central \tab w. nor. central \tab south atlantic \tab e. sou. central \tab w. sou. central \tab mountain \tab pacific \tab no answer \tab middle atlantic \tab Total \cr
   1990 \tab 1347 \tab 3 \tab 4 \tab 3 \tab 2 \tab 2 \tab 3 \tab 3 \tab 3 \tab 2 \tab - \tab 1372 \cr
   1991 \tab 1477 \tab 1 \tab 7 \tab 4 \tab 8 \tab 4 \tab 1 \tab 1 \tab 7 \tab 3 \tab 4 \tab 1517 \cr
   Total \tab 2824 \tab 4 \tab 11 \tab 7 \tab 10 \tab 6 \tab 4 \tab 4 \tab 10 \tab 5 \tab 4 \tab 2889 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{suiage}},
\code{\link{suiage2}},
\code{\link{suirac2}},
\code{\link{suirace}},
\code{\link{suireg}},
\code{\link{suisex2}},
\code{\link{suisex3}},
\code{\link{suiwho2}},
\code{\link{suiwho3}}
}
\concept{Victims & AIDS}
\keyword{variable}
