% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_551.R
\name{suireg3}
\alias{suireg3}
\title{Region of suicide victim 3rd closest to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
suireg3
}
\details{
Question We would like to know a few other things about that person.
D. What state did this person live in?
}
\section{Values}{

\itemize{
\item \verb{[0]} foreign
\item \verb{[1]} new england
\item \verb{[2]} middle atlantic
\item \verb{[3]} e. nor. central
\item \verb{[4]} w. nor. central
\item \verb{[5]} south atlantic
\item \verb{[6]} e. sou. central
\item \verb{[7]} w. sou. central
\item \verb{[8]} mountain
\item \verb{[9]} pacific
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5040/vshow}{GSS Data Explorer page for \code{suireg3}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab new england \tab e. nor. central \tab w. sou. central \tab pacific \tab no answer \tab middle atlantic \tab south atlantic \tab e. sou. central \tab mountain \tab Total \cr
   1990 \tab 1364 \tab 2 \tab 1 \tab 1 \tab 2 \tab 2 \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1502 \tab 1 \tab 1 \tab 1 \tab 2 \tab 3 \tab 2 \tab 2 \tab 2 \tab 1 \tab 1517 \cr
   Total \tab 2866 \tab 3 \tab 2 \tab 2 \tab 4 \tab 5 \tab 2 \tab 2 \tab 2 \tab 1 \tab 2889 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsads}},
\code{\link{aidshlth}},
\code{\link{aidsids}},
\code{\link{aidsinsr}},
\code{\link{aidsmar}},
\code{\link{aidssch}},
\code{\link{aidssxed}},
\code{\link{suiage3}},
\code{\link{suirac3}}
}
\concept{Victims & AIDS}
\keyword{variable}
