% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_73.R
\name{suirel2}
\alias{suirel2}
\title{R's relationship to 2nd closest suicide}
\source{
General Social Survey https://gss.norc.org
}
\description{
suirel2
}
\details{
Question Please tell me the relationship category of the person you knew second best that died by suicide.
(HAND CARD I4)
}
\section{Values}{

\itemize{
\item \code{1} husband or wife
\item \code{2} partner or lover
\item \code{3} son or daughter
\item \code{4} mother or father
\item \code{5} brother or sister
\item \code{6} step-brother or step-sister
\item \code{7} step-parent
\item \code{8} step-son or step-daughter
\item \code{9} cousin
\item \code{10} other relative
\item \code{11} friend
\item \code{12} neighbor
\item \code{13} co-worker
\item \code{14} patient
\item \code{15} acquaintance
\item \code{16} ex-spouse or ex-partner
\item \code{18} classmate
\item \code{19} relative of a friend
\item \code{20} student
\item \code{21} friend of a friend
\item \code{30} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6189/vshow}{GSS Data Explorer page for \code{suirel2}}.

Counts by year:\tabular{lllllllllllllllllllllll}{
   year \tab iap \tab husband or wife \tab partner or lover \tab son or daughter \tab mother or father \tab step-parent \tab cousin \tab other relative \tab friend \tab neighbor \tab co-worker \tab patient \tab acquaintance \tab ex-spouse or ex-partner \tab classmate \tab relative of a friend \tab student \tab friend of a friend \tab other \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2425 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 28 \tab 64 \tab 161 \tab 15 \tab 22 \tab 6 \tab 84 \tab 2 \tab 2 \tab 3 \tab 4 \tab 6 \tab 11 \tab 1 \tab 24 \tab 2867 \cr
   Total \tab 2425 \tab 1 \tab 4 \tab 1 \tab 2 \tab 1 \tab 28 \tab 64 \tab 161 \tab 15 \tab 22 \tab 6 \tab 84 \tab 2 \tab 2 \tab 3 \tab 4 \tab 6 \tab 11 \tab 1 \tab 24 \tab 2867 \cr
}
}

\seealso{
Other Suicide: 
\code{\link{suiclse1}},
\code{\link{suiclse2}},
\code{\link{suimths1}},
\code{\link{suimths2}},
\code{\link{suineg2}},
\code{\link{suitime1}},
\code{\link{suitime2}},
\code{\link{suiyrs1}},
\code{\link{suiyrs2}}
}
\concept{Suicide}
\keyword{variable}
