% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_73.R
\name{suitime1}
\alias{suitime1}
\title{How many years and months since closest suicide}
\source{
General Social Survey https://gss.norc.org
}
\description{
suitime1
}
\details{
Question How many months and years has it been since that person's death?
IF RESPONDENT PROVIDES AT LEAST A YEAR AMOUNT, PLEASE SELECT THE FIRST OPTION.
}
\section{Values}{

\itemize{
\item \code{1} respondent provides number of months or years
\item \code{2} respondent provides verbatim response uncodeable in months/years
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6186/vshow}{GSS Data Explorer page for \code{suitime1}}.
}

\seealso{
Other Suicide: 
\code{\link{suiclse1}},
\code{\link{suiclse2}},
\code{\link{suimths1}},
\code{\link{suimths2}},
\code{\link{suineg2}},
\code{\link{suirel2}},
\code{\link{suitime2}},
\code{\link{suiyrs1}},
\code{\link{suiyrs2}}
}
\concept{Suicide}
\keyword{variable}
