% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_73.R
\name{suitime2}
\alias{suitime2}
\title{How many years and months since 2nd closest suicide}
\source{
General Social Survey https://gss.norc.org
}
\description{
suitime2
}
\details{
Question How many months and years has it been since that person's death?
IF RESPONDENT PROVIDES AT LEAST A YEAR AMOUNT, PLEASE SELECT THE FIRST OPTION.
}
\section{Values}{

\itemize{
\item \code{1} respondent provides number of months or years
\item \code{2} respondent provides verbatim response uncodeable in months/years
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6192/vshow}{GSS Data Explorer page for \code{suitime2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab respondent provides number of months or years \tab respondent provides verbatim response uncodeable in months/years \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2425 \tab 384 \tab 23 \tab 10 \tab 25 \tab 2867 \cr
   Total \tab 2425 \tab 384 \tab 23 \tab 10 \tab 25 \tab 2867 \cr
}
}

\seealso{
Other Suicide: 
\code{\link{suiclse1}},
\code{\link{suiclse2}},
\code{\link{suimths1}},
\code{\link{suimths2}},
\code{\link{suineg2}},
\code{\link{suirel2}},
\code{\link{suitime1}},
\code{\link{suiyrs1}},
\code{\link{suiyrs2}}
}
\concept{Suicide}
\keyword{variable}
