% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_549.R
\name{suiwho}
\alias{suiwho}
\title{R's relationship to closest suicide}
\source{
General Social Survey https://gss.norc.org
}
\description{
suiwho
}
\details{
Question Think about the person you knew best that has committed suicide. Please tell me the letter of the category on the card which best describes your relationship to that person.
}
\section{Values}{

\itemize{
\item \code{1} husband or wife
\item \code{2} partner or lover
\item \code{3} son or daughter
\item \code{4} other relative
\item \code{5} friend
\item \code{6} neighbor
\item \code{7} co-worker
\item \code{8} acquaintance
\item \code{9} patient
\item \code{10} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990, 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5026/vshow}{GSS Data Explorer page for \code{suiwho}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab son or daughter \tab other relative \tab friend \tab neighbor \tab co-worker \tab acquaintance \tab other \tab no answer \tab partner or lover \tab patient \tab Total \cr
   1990 \tab 1238 \tab 1 \tab 18 \tab 45 \tab 9 \tab 10 \tab 43 \tab 6 \tab 2 \tab - \tab - \tab 1372 \cr
   1991 \tab 1344 \tab - \tab 24 \tab 62 \tab 10 \tab 15 \tab 46 \tab 5 \tab 4 \tab 2 \tab 5 \tab 1517 \cr
   Total \tab 2582 \tab 1 \tab 42 \tab 107 \tab 19 \tab 25 \tab 89 \tab 11 \tab 6 \tab 2 \tab 5 \tab 2889 \cr
}
}

\seealso{
Other Victims & AIDS: 
\code{\link{aidsage3}},
\code{\link{aidsded3}},
\code{\link{aidsrac3}},
\code{\link{aidsreg2}},
\code{\link{aidsreg3}},
\code{\link{aidssex3}},
\code{\link{aidswho3}},
\code{\link{suiknew}},
\code{\link{suisex}}
}
\concept{Victims & AIDS}
\keyword{variable}
