% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_73.R
\name{suiyrs2}
\alias{suiyrs2}
\title{How many years since 2nd closest suicide}
\source{
General Social Survey https://gss.norc.org
}
\description{
suiyrs2
}
\details{
Question YEARS:
}
\section{Values}{

\itemize{
\item \code{0} within the last year
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6194/vshow}{GSS Data Explorer page for \code{suiyrs2}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab within the last year \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 28 \tab 29 \tab 3 \tab 30 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 4 \tab 40 \tab 41 \tab 42 \tab 45 \tab 47 \tab 48 \tab 5 \tab 50 \tab 6 \tab 60 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2425 \tab 22 \tab 12 \tab 43 \tab 10 \tab 11 \tab 1 \tab 1 \tab 14 \tab 3 \tab 6 \tab 8 \tab 1 \tab 20 \tab 32 \tab 2 \tab 10 \tab 6 \tab 3 \tab 13 \tab 2 \tab 1 \tab 2 \tab 30 \tab 15 \tab 1 \tab 2 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 15 \tab 8 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 34 \tab 5 \tab 15 \tab 2 \tab 11 \tab 14 \tab 6 \tab 5 \tab 39 \tab 2867 \cr
   Total \tab 2425 \tab 22 \tab 12 \tab 43 \tab 10 \tab 11 \tab 1 \tab 1 \tab 14 \tab 3 \tab 6 \tab 8 \tab 1 \tab 20 \tab 32 \tab 2 \tab 10 \tab 6 \tab 3 \tab 13 \tab 2 \tab 1 \tab 2 \tab 30 \tab 15 \tab 1 \tab 2 \tab 1 \tab 4 \tab 1 \tab 1 \tab 2 \tab 15 \tab 8 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 34 \tab 5 \tab 15 \tab 2 \tab 11 \tab 14 \tab 6 \tab 5 \tab 39 \tab 2867 \cr
}
}

\seealso{
Other Suicide: 
\code{\link{suiclse1}},
\code{\link{suiclse2}},
\code{\link{suimths1}},
\code{\link{suimths2}},
\code{\link{suineg2}},
\code{\link{suirel2}},
\code{\link{suitime1}},
\code{\link{suitime2}},
\code{\link{suiyrs1}}
}
\concept{Suicide}
\keyword{variable}
