% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_115.R
\name{sunsch16}
\alias{sunsch16}
\title{How frequently r attended sunday school}
\source{
General Social Survey https://gss.norc.org
}
\description{
sunsch16
}
\details{
Question When you were growing up, did you attend Sunday school or religious instruction classes regularly, most of the time, some of the time, or never?
}
\section{Values}{

\itemize{
\item \code{1} regularly
\item \code{2} most of the time
\item \code{3} some of the time
\item \code{4} never
\item \code{5} att parochial sch
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1069/vshow}{GSS Data Explorer page for \code{sunsch16}}.
}

\seealso{
Other Religion: 
\code{\link{churhsch}},
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{maoth}},
\code{\link{paden}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{paoth}},
\code{\link{parelig}}
}
\concept{Religion}
\keyword{variable}
