% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_326.R
\name{supfam}
\alias{supfam}
\title{Young should be able to support family}
\source{
General Social Survey https://gss.norc.org
}
\description{
supfam
}
\details{
Question People differ in their ideas about what it takes for a young person to become an adult these days. How important is it for them to be...
Be capable of supporting a family financially
}
\section{Values}{

\itemize{
\item \verb{[1]} extremely important
\item \verb{[2]} quite important
\item \verb{[3]} somewhat important
\item \verb{[4]} not too important
\item \verb{[5]} not at all important
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2904/vshow}{GSS Data Explorer page for \code{supfam}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab extremely important \tab quite important \tab somewhat important \tab not too important \tab not at all important \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 824 \tab 306 \tab 152 \tab 57 \tab 27 \tab 21 \tab 6 \tab 2765 \cr
   Total \tab 1372 \tab 824 \tab 306 \tab 152 \tab 57 \tab 27 \tab 21 \tab 6 \tab 2765 \cr
}
}

\seealso{
Other Adult Transitions: 
\code{\link{boycott}},
\code{\link{conoffcl}},
\code{\link{getmar}},
\code{\link{getmar1}},
\code{\link{havchld}},
\code{\link{havchld1}},
\code{\link{protest}},
\code{\link{signpet}},
\code{\link{supfam1}}
}
\concept{Adult Transitions}
\keyword{variable}
