% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_502.R
\name{suprvsjb}
\alias{suprvsjb}
\title{Does r supervise others at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
suprvsjb
}
\details{
Question Do you supervise others at work as a part of your job?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6028/vshow}{GSS Data Explorer page for \code{suprvsjb}}.
}

\seealso{
Other Work Orientation: 
\code{\link{employer}},
\code{\link{localnum}},
\code{\link{moonlite}},
\code{\link{numemply}},
\code{\link{numsites}},
\code{\link{othhrs}},
\code{\link{supervis}},
\code{\link{supnum}},
\code{\link{totalnum}}
}
\concept{Work Orientation}
\keyword{variable}
