% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_502.R
\name{suprvsjb}
\alias{suprvsjb}
\title{Does r supervise others at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
suprvsjb
}
\details{
Question Do you supervise others at work as a part of your job?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2010, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6028/vshow}{GSS Data Explorer page for \code{suprvsjb}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2010 \tab 857 \tab 419 \tab 744 \tab 1 \tab 23 \tab 2044 \cr
   2014 \tab 1288 \tab 455 \tab 790 \tab - \tab 5 \tab 2538 \cr
   Total \tab 2145 \tab 874 \tab 1534 \tab 1 \tab 28 \tab 4582 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{employer}},
\code{\link{localnum}},
\code{\link{moonlite}},
\code{\link{numemply}},
\code{\link{numsites}},
\code{\link{othhrs}},
\code{\link{supervis}},
\code{\link{supnum}},
\code{\link{totalnum}}
}
\concept{Work Orientation}
\keyword{variable}
