% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_653.R
\name{svyenjoy}
\alias{svyenjoy}
\title{Enjoy taking surveys}
\source{
General Social Survey https://gss.norc.org
}
\description{
svyenjoy
}
\details{
Question Overall, how much do you enjoy participating in surveys? Would you say a great deal, a good bit, some, a little, or not at all?
}
\section{Values}{

\itemize{
\item \code{1} a great deal
\item \code{2} a good bit
\item \code{3} some
\item \code{4} a little
\item \code{5} not at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7863/vshow}{GSS Data Explorer page for \code{svyenjoy}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab a great deal \tab a good bit \tab some \tab a little \tab not at all \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 66 \tab 508 \tab 854 \tab 1074 \tab 600 \tab 404 \tab 9 \tab 21 \tab 8 \tab 3544 \cr
   Total \tab 66 \tab 508 \tab 854 \tab 1074 \tab 600 \tab 404 \tab 9 \tab 21 \tab 8 \tab 3544 \cr
}
}

\keyword{variable}
