% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_654.R
\name{svyid1}
\alias{svyid1}
\title{Willingness to use video chat}
\source{
General Social Survey https://gss.norc.org
}
\description{
svyid1
}
\details{
Question How likely is it that you would be willing to use a video chat for a survey interview within the next few months?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7864/vshow}{GSS Data Explorer page for \code{svyid1}}.
}

\keyword{variable}
