% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_111.R
\name{swayvote}
\alias{swayvote}
\title{Has r tried to influence votes of others}
\source{
General Social Survey https://gss.norc.org
}
\description{
swayvote
}
\details{
Question During elections, do you ever try to show people why they should vote for one of the parties or candidates? Do you do that often, sometimes, rarely or never?
}
\section{Values}{

\itemize{
\item \verb{[1]} often
\item \verb{[2]} sometimes
\item \verb{[3]} rarely
\item \verb{[4]} never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1028/vshow}{GSS Data Explorer page for \code{swayvote}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab often \tab sometimes \tab rarely \tab never \tab don't know \tab no answer \tab Total \cr
   1987 \tab - \tab 175 \tab 398 \tab 247 \tab 986 \tab 4 \tab 9 \tab 1819 \cr
   Total \tab 0 \tab 175 \tab 398 \tab 247 \tab 986 \tab 4 \tab 9 \tab 1819 \cr
}
}

\seealso{
Other Socio-political participation: 
\code{\link{intpol}},
\code{\link{lobbygrp}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{locinflu}},
\code{\link{loclived}},
\code{\link{locprob}},
\code{\link{loctrust}},
\code{\link{workpol}}
}
\concept{Socio-political participation}
\keyword{variable}
