% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_114.R
\name{switch1}
\alias{switch1}
\title{What was r's first religious preference}
\source{
General Social Survey https://gss.norc.org
}
\description{
switch1
}
\details{
Question I'd like to go over your religious preferences since you were raised as a (RELIGION AND/OR DENOMINATION MENTIONED IN RELIG16/ DENOM16).
Have you ever had another religious preference besides being (RELIGION AND/OR CHURCH DENOMINATION menTIONED IN RELIG16/DENOM16).
Yes................(ASK A)
No..............(GO TO MARELIG)
A. Starting when you were 16 years old, what other religious preferences have you had?  Please mention your earliest preference first. Any others?
\enumerate{
\item First Mention
}
}
\section{Values}{

\itemize{
\item \code{110000} amer baptist assoc
\item \code{111000} amer bapt ch in us
\item \code{112000} natl bapt con of am
\item \code{113000} natl bapt con, usa
\item \code{114000} southern bapt con
\item \code{115000} other baptist church
\item \code{118000} baptist, dk which
\item \code{120000} afrcn meth epis ch
\item \code{121000} afrcn meth epis zion
\item \code{122000} united methodist ch
\item \code{123000} other methodist ch
\item \code{128000} methodist, dk which
\item \code{130000} amer lutheran church
\item \code{131000} lutheran ch in amer
\item \code{132000} luth ch-missouri syn
\item \code{133000} wisconsin ev luth sn
\item \code{134000} other lutheran churc
\item \code{135000} evangelical lutheran
\item \code{138000} lutheran, dk which
\item \code{140000} presbyterian ch n us
\item \code{141000} united pres ch in us
\item \code{142000} other presbyterian
\item \code{148000} presbyterian, dk whi
\item \code{150000} episcopal church
\item \code{160000} other protestant
\item \code{170000} no denomination
\item \code{2e+05} catholic
\item \code{3e+05} jewish
\item \code{4e+05} none
\item \code{5e+05} other religion
\item \code{999999} missing
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1050/vshow}{GSS Data Explorer page for \code{switch1}}.
}

\seealso{
Other Religion: 
\code{\link{believe}},
\code{\link{blkwhite}},
\code{\link{bwfriend}},
\code{\link{cathsch}},
\code{\link{chrissch}},
\code{\link{churhact}},
\code{\link{churhmem}},
\code{\link{churhsch}},
\code{\link{decbible}},
\code{\link{decchurh}},
\code{\link{decoths}},
\code{\link{decself}},
\code{\link{denomsch}},
\code{\link{doubts1}},
\code{\link{doubts2}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{faith1}},
\code{\link{faith2}},
\code{\link{faith3}},
\code{\link{faith4}},
\code{\link{ferespct}},
\code{\link{follow}},
\code{\link{friends}},
\code{\link{frndcon1}},
\code{\link{frndcon2}},
\code{\link{frndcon3}},
\code{\link{frndcon4}},
\code{\link{frndcon5}},
\code{\link{frndden1}},
\code{\link{frndden2}},
\code{\link{frndden3}},
\code{\link{frndden4}},
\code{\link{frndden5}},
\code{\link{frndfnd1}},
\code{\link{frndfnd2}},
\code{\link{frndfnd3}},
\code{\link{frndfnd4}},
\code{\link{frndfnd5}},
\code{\link{frndoth1}},
\code{\link{frndoth2}},
\code{\link{frndoth3}},
\code{\link{frndoth4}},
\code{\link{frndoth5}},
\code{\link{frndrac1}},
\code{\link{frndrac2}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{frndrel1}},
\code{\link{frndrel2}},
\code{\link{frndrel3}},
\code{\link{frndrel4}},
\code{\link{frndrel5}},
\code{\link{givearts}},
\code{\link{givecong}},
\code{\link{giveoth}},
\code{\link{giverel}},
\code{\link{gochurch}},
\code{\link{god}},
\code{\link{godsells}},
\code{\link{godsport}},
\code{\link{goownway}},
\code{\link{grace16}},
\code{\link{homesch}},
\code{\link{implives}},
\code{\link{join16}},
\code{\link{kid5up}},
\code{\link{lapsed}},
\code{\link{madatgod}},
\code{\link{maden}},
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{maoth}},
\code{\link{marelig}},
\code{\link{myfaith}},
\code{\link{numbwfrd}},
\code{\link{numcong}},
\code{\link{numfrend}},
\code{\link{obeytch}},
\code{\link{othchrch}},
\code{\link{othsch}},
\code{\link{paden}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{paoth}},
\code{\link{parelig}},
\code{\link{permoral}},
\code{\link{preach}},
\code{\link{privsch}},
\code{\link{pubsch}},
\code{\link{punsin}},
\code{\link{readword}},
\code{\link{reborn}},
\code{\link{relhrs1}},
\code{\link{relhrs2}},
\code{\link{relsch}},
\code{\link{rotapple}},
\code{\link{savesoul}},
\code{\link{saygrace}},
\code{\link{scichng}},
\code{\link{scimoral}},
\code{\link{scipry}},
\code{\link{scisolve}},
\code{\link{shopmove}},
\code{\link{shopnum}},
\code{\link{shoprel}},
\code{\link{sunsch16}},
\code{\link{switage1}},
\code{\link{switage2}},
\code{\link{switch2}},
\code{\link{switch3}},
\code{\link{switched}},
\code{\link{switnum}},
\code{\link{switwhy1}},
\code{\link{switwhy2}},
\code{\link{sympcoun}},
\code{\link{tvrelig}},
\code{\link{worship}},
\code{\link{wrkyoung}}
}
\concept{Religion}
\keyword{variable}
