% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_114.R
\name{switwhy1}
\alias{switwhy1}
\title{Why did r make first switch}
\source{
General Social Survey https://gss.norc.org
}
\description{
switwhy1
}
\details{
Question I'd like to go over your religious preferences since you were raised as a (RELIGION AND/OR DENOMINATION MENTIONED IN RELIG16/DENOM16).
Have you ever had another religious preference besides being (RELIGION AND/OR CHURCH DENOMINATION MENTIONED IN RELIG16/DENOM16).
Yes................(ASK A)
No..............(GO TO MARELIG)
C2. Why did you become a (FIRST/SECOND MENTION)?
(1) Why first switch made
}
\section{Values}{

\itemize{
\item \code{0} na
\item \code{10} marriage
\item \code{14} marr & location
\item \code{15} marr & theology
\item \code{16} marr. & clergy
\item \code{17} marr & positive blf
\item \code{20} friends
\item \code{23} frnds & family
\item \code{24} frnds & location
\item \code{25} frnds & theo dis
\item \code{26} frnds & clergy
\item \code{28} frnds & other pos
\item \code{30} family
\item \code{34} fam & location
\item \code{35} fam & theo diss
\item \code{37} fam & positive blf
\item \code{39} fam & misc
\item \code{40} location
\item \code{41} loc & to change
\item \code{45} loc & theo dissat
\item \code{46} loc and clergy
\item \code{47} loc & positive blf
\item \code{50} theological diss
\item \code{56} theo diss & clergy
\item \code{60} clergy
\item \code{68} clergy & oth pos
\item \code{69} clergy & missing
\item \code{70} to change
\item \code{71} other, to change
\item \code{76} other, misc
\item \code{77} other, pos blfs
\item \code{78} other, pos other
\item \code{79} other, missing
\item \code{85} loc, theo & clrgy
\item \code{86} frnds,fam,location
\item \code{87} marr,loc,theo dis
\item \code{99} missing values
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1056/vshow}{GSS Data Explorer page for \code{switwhy1}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllll}{
   year \tab iap \tab marriage \tab marr & location \tab marr & theology \tab marr. & clergy \tab marr & positive blf \tab friends \tab frnds & family \tab frnds & location \tab frnds & clergy \tab frnds & other pos \tab family \tab fam & location \tab fam & theo diss \tab fam & positive blf \tab fam & misc \tab location \tab loc & theo dissat \tab loc and clergy \tab loc & positive blf \tab theological diss \tab theo diss & clergy \tab clergy \tab clergy & missing \tab other, to change \tab other, misc \tab other, pos blfs \tab other, pos other \tab other, missing \tab no answer \tab Total \cr
   1988 \tab 945 \tab 71 \tab 1 \tab 2 \tab 1 \tab 3 \tab 47 \tab 2 \tab 4 \tab 2 \tab 1 \tab 46 \tab 1 \tab 2 \tab 1 \tab 1 \tab 30 \tab 1 \tab 2 \tab 3 \tab 63 \tab 1 \tab 8 \tab 1 \tab 3 \tab 31 \tab 14 \tab 8 \tab 1 \tab 185 \tab 1481 \cr
   Total \tab 945 \tab 71 \tab 1 \tab 2 \tab 1 \tab 3 \tab 47 \tab 2 \tab 4 \tab 2 \tab 1 \tab 46 \tab 1 \tab 2 \tab 1 \tab 1 \tab 30 \tab 1 \tab 2 \tab 3 \tab 63 \tab 1 \tab 8 \tab 1 \tab 3 \tab 31 \tab 14 \tab 8 \tab 1 \tab 185 \tab 1481 \cr
}
}

\seealso{
Other Religion: 
\code{\link{maden}},
\code{\link{marelig}},
\code{\link{switage1}},
\code{\link{switage2}},
\code{\link{switch1}},
\code{\link{switch2}},
\code{\link{switch3}},
\code{\link{switnum}},
\code{\link{switwhy2}}
}
\concept{Religion}
\keyword{variable}
