% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_114.R
\name{switwhy2}
\alias{switwhy2}
\title{Why did r make second switch}
\source{
General Social Survey https://gss.norc.org
}
\description{
switwhy2
}
\details{
Question I'd like to go over your religious preferences since you were raised as a (RELIGION AND/OR DENOMINATION MENTIONED IN RELIG16/DENOM16).
Have you ever had another religious preference besides being (RELIGION AND/OR CHURCH DENOMINATION MENTIONED IN RELIG16/DENOM16).
Yes................(ASK A)
No..............(GO TO MARELIG)
C2. Why did you become a (FIRST/SECOND MENTION)?
(2) Why second switch made
}
\section{Values}{

\itemize{
\item \code{0} na
\item \code{10} marriage
\item \code{14} marr & location
\item \code{15} marr & theology
\item \code{16} marr. & clergy
\item \code{17} marr & positive blf
\item \code{20} friends
\item \code{23} frnds & family
\item \code{24} frnds & location
\item \code{25} frnds & theo dis
\item \code{26} frnds & clergy
\item \code{28} frnds & other pos
\item \code{30} family
\item \code{34} fam & location
\item \code{35} fam & theo diss
\item \code{37} fam & positive blf
\item \code{39} fam & misc
\item \code{40} location
\item \code{41} loc & to change
\item \code{45} loc & theo dissat
\item \code{46} loc and clergy
\item \code{47} loc & positive blf
\item \code{50} theological diss
\item \code{56} theo diss & clergy
\item \code{60} clergy
\item \code{68} clergy & oth pos
\item \code{69} clergy & missing
\item \code{70} to change
\item \code{71} other, to change
\item \code{76} other, misc
\item \code{77} other, pos blfs
\item \code{78} other, pos other
\item \code{79} other, missing
\item \code{85} loc, theo & clrgy
\item \code{86} frnds,fam,location
\item \code{87} marr,loc,theo dis
\item \code{99} missing values
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1057/vshow}{GSS Data Explorer page for \code{switwhy2}}.
}

\seealso{
Other Religion: 
\code{\link{believe}},
\code{\link{blkwhite}},
\code{\link{bwfriend}},
\code{\link{cathsch}},
\code{\link{chrissch}},
\code{\link{churhact}},
\code{\link{churhmem}},
\code{\link{churhsch}},
\code{\link{decbible}},
\code{\link{decchurh}},
\code{\link{decoths}},
\code{\link{decself}},
\code{\link{denomsch}},
\code{\link{doubts1}},
\code{\link{doubts2}},
\code{\link{doubts3}},
\code{\link{doubts4}},
\code{\link{faith1}},
\code{\link{faith2}},
\code{\link{faith3}},
\code{\link{faith4}},
\code{\link{ferespct}},
\code{\link{follow}},
\code{\link{friends}},
\code{\link{frndcon1}},
\code{\link{frndcon2}},
\code{\link{frndcon3}},
\code{\link{frndcon4}},
\code{\link{frndcon5}},
\code{\link{frndden1}},
\code{\link{frndden2}},
\code{\link{frndden3}},
\code{\link{frndden4}},
\code{\link{frndden5}},
\code{\link{frndfnd1}},
\code{\link{frndfnd2}},
\code{\link{frndfnd3}},
\code{\link{frndfnd4}},
\code{\link{frndfnd5}},
\code{\link{frndoth1}},
\code{\link{frndoth2}},
\code{\link{frndoth3}},
\code{\link{frndoth4}},
\code{\link{frndoth5}},
\code{\link{frndrac1}},
\code{\link{frndrac2}},
\code{\link{frndrac3}},
\code{\link{frndrac4}},
\code{\link{frndrac5}},
\code{\link{frndrel1}},
\code{\link{frndrel2}},
\code{\link{frndrel3}},
\code{\link{frndrel4}},
\code{\link{frndrel5}},
\code{\link{givearts}},
\code{\link{givecong}},
\code{\link{giveoth}},
\code{\link{giverel}},
\code{\link{gochurch}},
\code{\link{god}},
\code{\link{godsells}},
\code{\link{godsport}},
\code{\link{goownway}},
\code{\link{grace16}},
\code{\link{homesch}},
\code{\link{implives}},
\code{\link{join16}},
\code{\link{kid5up}},
\code{\link{lapsed}},
\code{\link{madatgod}},
\code{\link{maden}},
\code{\link{mafund}},
\code{\link{majew}},
\code{\link{maoth}},
\code{\link{marelig}},
\code{\link{myfaith}},
\code{\link{numbwfrd}},
\code{\link{numcong}},
\code{\link{numfrend}},
\code{\link{obeytch}},
\code{\link{othchrch}},
\code{\link{othsch}},
\code{\link{paden}},
\code{\link{pafund}},
\code{\link{pajew}},
\code{\link{paoth}},
\code{\link{parelig}},
\code{\link{permoral}},
\code{\link{preach}},
\code{\link{privsch}},
\code{\link{pubsch}},
\code{\link{punsin}},
\code{\link{readword}},
\code{\link{reborn}},
\code{\link{relhrs1}},
\code{\link{relhrs2}},
\code{\link{relsch}},
\code{\link{rotapple}},
\code{\link{savesoul}},
\code{\link{saygrace}},
\code{\link{scichng}},
\code{\link{scimoral}},
\code{\link{scipry}},
\code{\link{scisolve}},
\code{\link{shopmove}},
\code{\link{shopnum}},
\code{\link{shoprel}},
\code{\link{sunsch16}},
\code{\link{switage1}},
\code{\link{switage2}},
\code{\link{switch1}},
\code{\link{switch2}},
\code{\link{switch3}},
\code{\link{switched}},
\code{\link{switnum}},
\code{\link{switwhy1}},
\code{\link{sympcoun}},
\code{\link{tvrelig}},
\code{\link{worship}},
\code{\link{wrkyoung}}
}
\concept{Religion}
\keyword{variable}
