% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_125.R
\name{sympcoun}
\alias{sympcoun}
\title{How well local church counsels sympathetically}
\source{
General Social Survey https://gss.norc.org
}
\description{
sympcoun
}
\details{
Question How do you rate how your local church does each of the following?  Would you say your local church does an excellent, very good, fair, or poor job of:
C. Sympathetic counseling
}
\section{Values}{

\itemize{
\item \code{1} excellent
\item \code{2} very good
\item \code{3} good
\item \code{4} fair
\item \code{5} poor
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1163/vshow}{GSS Data Explorer page for \code{sympcoun}}.
}

\seealso{
Other Religion: 
\code{\link{ferespct}},
\code{\link{obeytch}},
\code{\link{preach}},
\code{\link{worship}},
\code{\link{wrkyoung}}

Other Religion: 
\code{\link{ferespct}},
\code{\link{obeytch}},
\code{\link{preach}},
\code{\link{worship}},
\code{\link{wrkyoung}}
}
\concept{Religion}
\keyword{variable}
