% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_172.R
\name{symptblk}
\alias{symptblk}
\title{R feel sympathy towards blacks}
\source{
General Social Survey https://gss.norc.org
}
\description{
symptblk
}
\details{
Question Now, I would like to ask whether you have ever felt the following ways about blacks and their families. For each of the feelings that I ask you about, please tell me whether you have felt that way very often, fairly often, not too often, or never?
A. How often have you felt sympathy for blacks?
}
\section{Values}{

\itemize{
\item \code{1} very often
\item \code{2} fairly often
\item \code{3} not too often
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1605/vshow}{GSS Data Explorer page for \code{symptblk}}.
}

\seealso{
Other Multiculturalism: 
\code{\link{admirblk}},
\code{\link{bilinged}},
\code{\link{engballt}},
\code{\link{engoffcl}},
\code{\link{engteach}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}},
\code{\link{letin}}

Other Blacks: 
\code{\link{admirblk}}

Other Race: 
\code{\link{admirblk}},
\code{\link{ethorgs}},
\code{\link{ethspkno}},
\code{\link{ethspkok}}
}
\concept{Blacks}
\concept{Multiculturalism}
\concept{Race}
\keyword{variable}
