% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_198.R
\name{tablprce}
\alias{tablprce}
\title{The factory not lowering price acceptable?}
\source{
General Social Survey https://gss.norc.org
}
\description{
tablprce
}
\details{
Question Consider the following situation: A small factory produces tables and sells all that it can make at $200 each. because of changes in the price of materials, the cost of making each table has recently decreased by $20. The factory does not lower its price for the tables. Is this acceptable or unfair?
}
\section{Values}{

\itemize{
\item \verb{[1]} acceptable
\item \verb{[2]} unfair
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1865/vshow}{GSS Data Explorer page for \code{tablprce}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab acceptable \tab unfair \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 737 \tab 662 \tab 40 \tab 5 \tab 2904 \cr
   Total \tab 1460 \tab 737 \tab 662 \tab 40 \tab 5 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{adoption}},
\code{\link{frndawk}},
\code{\link{frnddeal}},
\code{\link{organsb}},
\code{\link{organsw}},
\code{\link{sellbaby}},
\code{\link{sellorgn}},
\code{\link{sellsex}},
\code{\link{transrel}}
}
\concept{Markets}
\keyword{variable}
