% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_199.R
\name{tagsales}
\alias{tagsales}
\title{How frequently do you go to garage sales etc}
\source{
General Social Survey https://gss.norc.org
}
\description{
tagsales
}
\details{
Question How frequently do you go to garage sales, tag sales, or yard sales -- that is, one-time sales in which homeowners sell items that they no longer need from out of their homes, yards, or garages?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} rarely
\item \code{3} occasionally 1-3 times a yr
\item \code{4} regularly
\item \code{5} almost every weekend
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1874/vshow}{GSS Data Explorer page for \code{tagsales}}.
}

\seealso{
Other Markets: 
\code{\link{decauto}},
\code{\link{econsys}},
\code{\link{fambiz}},
\code{\link{famfinan}},
\code{\link{haggle}},
\code{\link{hrfambiz}},
\code{\link{natrecon}},
\code{\link{reqinfo}},
\code{\link{wkfambiz}}
}
\concept{Markets}
\keyword{variable}
