% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_125.R
\name{talkback}
\alias{talkback}
\title{Talking back to parents is bad}
\source{
General Social Survey https://gss.norc.org
}
\description{
talkback
}
\details{
Question Now I am going to read you several statements. For each one please tell me whether you agree strongly, agree, disagree, or disagree strongly.
B. A child should never be allowed to talk back to his parents, or else he will lose respect for them
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1167/vshow}{GSS Data Explorer page for \code{talkback}}.
}

\seealso{
Other Intergroup Relations: 
\code{\link{openmind}},
\code{\link{ownthing}},
\code{\link{twoclass}}
}
\concept{Intergroup Relations}
\keyword{variable}
