% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_322.R
\name{talkemp}
\alias{talkemp}
\title{How likely to talk directly if coworker not work}
\source{
General Social Survey https://gss.norc.org
}
\description{
talkemp
}
\details{
Question If you were to see a fellow employee not working as hard or well as he or she should, how likely would you be to:
A. Talk directly to the employee
}
\section{Values}{

\itemize{
\item \code{1} not at all likely
\item \code{2} not very likely
\item \code{3} somewhat likely
\item \code{4} very likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2865/vshow}{GSS Data Explorer page for \code{talkemp}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab not at all likely \tab not very likely \tab somewhat likely \tab very likely \tab don't know \tab no answer \tab Total \cr
   2002 \tab 969 \tab 480 \tab 301 \tab 432 \tab 514 \tab 40 \tab 29 \tab 2765 \cr
   2006 \tab 2778 \tab 420 \tab 284 \tab 383 \tab 546 \tab 69 \tab 30 \tab 4510 \cr
   Total \tab 3747 \tab 900 \tab 585 \tab 815 \tab 1060 \tab 109 \tab 59 \tab 7275 \cr
}
}

\seealso{
Other Employee Compensation: 
\code{\link{cowrkhrd}},
\code{\link{donothng}},
\code{\link{lastwkmo}},
\code{\link{lastwkyr}},
\code{\link{talksup}}
}
\concept{Employee Compensation}
\keyword{variable}
