% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_105.R
\name{talkpol}
\alias{talkpol}
\title{How often does r discuss politics?}
\source{
General Social Survey https://gss.norc.org
}
\description{
talkpol
}
\details{
Question Thinking of the (people/person) we''ve just been discussing, would you say you discuss social and political issues with (him/her/them) almost all the time, most of the time, occasionally, or almost never?
}
\section{Values}{

\itemize{
\item \verb{[1]} almost always
\item \verb{[2]} most of the time
\item \verb{[3]} occasionally
\item \verb{[4]} almost never
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/961/vshow}{GSS Data Explorer page for \code{talkpol}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab almost always \tab most of the time \tab occasionally \tab almost never \tab don't know \tab no answer \tab Total \cr
   1985 \tab 136 \tab 246 \tab 257 \tab 724 \tab 143 \tab 3 \tab 25 \tab 1534 \cr
   Total \tab 136 \tab 246 \tab 257 \tab 724 \tab 143 \tab 3 \tab 25 \tab 1534 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{frndknow}},
\code{\link{talkpol3}}
}
\concept{Social Networks}
\keyword{variable}
