% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_105.R
\name{talkpol3}
\alias{talkpol3}
\title{R discusses politics with person #3}
\source{
General Social Survey https://gss.norc.org
}
\description{
talkpol3
}
\details{
Question About how often do you talk to (NAME) about political matters?
3. Political Talks with Third Person
}
\section{Values}{

\itemize{
\item \code{1} almost daily
\item \code{2} at least weekly
\item \code{3} at least monthly
\item \code{4} at least yearly
\item \code{5} less than yearly
\item \code{6} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/960/vshow}{GSS Data Explorer page for \code{talkpol3}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab almost daily \tab at least weekly \tab at least monthly \tab at least yearly \tab less than yearly \tab never \tab don't know \tab no answer \tab Total \cr
   1987 \tab 735 \tab 78 \tab 165 \tab 257 \tab 237 \tab 108 \tab 207 \tab 4 \tab 28 \tab 1819 \cr
   Total \tab 735 \tab 78 \tab 165 \tab 257 \tab 237 \tab 108 \tab 207 \tab 4 \tab 28 \tab 1819 \cr
}
}

\seealso{
Other Social Networks: 
\code{\link{frndknow}},
\code{\link{talkpol}}
}
\concept{Social Networks}
\keyword{variable}
