% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_601.R
\name{talkspvs}
\alias{talkspvs}
\title{Comfortable talking with supervisor about personal}
\source{
General Social Survey https://gss.norc.org
}
\description{
talkspvs
}
\details{
Question How comfortable are you talking with your supervisor about conflicts between work and personal or family life
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all comfortable
\item \verb{[2]} a little
\item \verb{[3]} somewhat
\item \verb{[4]} very
\item \verb{[5]} extremely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5505/vshow}{GSS Data Explorer page for \code{talkspvs}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab not at all comfortable \tab a little \tab somewhat \tab very \tab extremely \tab don't know \tab no answer \tab Total \cr
   2014 \tab 1505 \tab 139 \tab 119 \tab 262 \tab 337 \tab 166 \tab 5 \tab 5 \tab 2538 \cr
   Total \tab 1505 \tab 139 \tab 119 \tab 262 \tab 337 \tab 166 \tab 5 \tab 5 \tab 2538 \cr
}
}

\seealso{
Other Work and Relationships: 
\code{\link{effctsup}},
\code{\link{wkcontct}}
}
\concept{Work and Relationships}
\keyword{variable}
