% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_391.R
\name{talkteam}
\alias{talkteam}
\title{If employee not working: talk directly to the employee}
\source{
General Social Survey https://gss.norc.org
}
\description{
talkteam
}
\details{
Question If you were to see a fellow employee not working as hard or well as he or she should, how likely would you be to talk about it in a work group or team? (Would you say…)
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all likely
\item \verb{[2]} not very likely
\item \verb{[3]} somewhat likely
\item \verb{[4]} very likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3555/vshow}{GSS Data Explorer page for \code{talkteam}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab not at all likely \tab not very likely \tab somewhat likely \tab very likely \tab don't know \tab no answer \tab Total \cr
   2006 \tab 2778 \tab 575 \tab 305 \tab 424 \tab 325 \tab 68 \tab 35 \tab 4510 \cr
   Total \tab 2778 \tab 575 \tab 305 \tab 424 \tab 325 \tab 68 \tab 35 \tab 4510 \cr
}
}

\seealso{
Other Shared Capitalism: 
\code{\link{compwage}},
\code{\link{empinput}},
\code{\link{emptrain}},
\code{\link{esop}},
\code{\link{extr2005}},
\code{\link{numorg}},
\code{\link{ops2005}},
\code{\link{slfmangd}},
\code{\link{wealth}}
}
\concept{Shared Capitalism}
\keyword{variable}
