% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_391.R
\name{talkteam}
\alias{talkteam}
\title{If employee not working: talk directly to the employee}
\source{
General Social Survey https://gss.norc.org
}
\description{
talkteam
}
\details{
Question If you were to see a fellow employee not working as hard or well as he or she should, how likely would you be to talk about it in a work group or team? (Would you say…)
}
\section{Values}{

\itemize{
\item \code{1} not at all likely
\item \code{2} not very likely
\item \code{3} somewhat likely
\item \code{4} very likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3555/vshow}{GSS Data Explorer page for \code{talkteam}}.
}

\seealso{
Other Shared Capitalism: 
\code{\link{compwage}},
\code{\link{empinput}},
\code{\link{emptrain}},
\code{\link{esop}},
\code{\link{extr2005}},
\code{\link{numorg}},
\code{\link{ops2005}},
\code{\link{slfmangd}},
\code{\link{wealth}}
}
\concept{Shared Capitalism}
\keyword{variable}
