% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_102.R
\name{talkto5}
\alias{talkto5}
\title{How often does r talk to person #5?}
\source{
General Social Survey https://gss.norc.org
}
\description{
talkto5
}
\details{
Question Thinking about how often you usually talk to (NAME), on the average, Do you talk to (him/her) almost every day, at least Once a week, at least once a month, or less than once a month? ASK FOR EACH NAME.
5. Talks to Fifth Person
}
\section{Values}{

\itemize{
\item \code{1} almost daily
\item \code{2} once a week
\item \code{3} once a month
\item \code{4} lt once a month
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/- \tab full \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/934/vshow}{GSS Data Explorer page for \code{talkto5}}.
}

\seealso{
Other Social Networks: 
\code{\link{known1}},
\code{\link{known2}},
\code{\link{known3}},
\code{\link{known4}},
\code{\link{known5}},
\code{\link{talkto1}},
\code{\link{talkto2}},
\code{\link{talkto3}},
\code{\link{talkto4}}

Other Social Networks: 
\code{\link{known1}},
\code{\link{known2}},
\code{\link{known3}},
\code{\link{known4}},
\code{\link{known5}},
\code{\link{talkto1}},
\code{\link{talkto2}},
\code{\link{talkto3}},
\code{\link{talkto4}}
}
\concept{Social Networks}
\keyword{variable}
