% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_449.R
\name{tapphone}
\alias{tapphone}
\title{Should authorities have right to tap phone conver'sation}
\source{
General Social Survey https://gss.norc.org
}
\description{
tapphone
}
\details{
Question Suppose the government suspected that a terrorist act was about to happen. Do you think the authorities should have the right to…
B. Tap people’s telephone conversations?
}
\section{Values}{

\itemize{
\item \code{1} definitely should have right
\item \code{2} probably should have right
\item \code{3} probably should not have right
\item \code{4} definitely should not have right
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2016 \cr
   D/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4115/vshow}{GSS Data Explorer page for \code{tapphone}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should have right \tab probably should have right \tab probably should not have right \tab definitely should not have right \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 2992 \tab 380 \tab 435 \tab 280 \tab 405 \tab 18 \tab - \tab 4510 \cr
   2016 \tab 1477 \tab 309 \tab 389 \tab 285 \tab 377 \tab 17 \tab 13 \tab 2867 \cr
   Total \tab 4469 \tab 689 \tab 824 \tab 565 \tab 782 \tab 35 \tab 13 \tab 7377 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{exploit}},
\code{\link{fewtrsty}},
\code{\link{givinffor}},
\code{\link{givinfusa}},
\code{\link{hlpinflu}},
\code{\link{polsfair}},
\code{\link{stoprndm}},
\code{\link{wotrial}},
\code{\link{youinflu}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
