% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_30.R
\name{tax}
\alias{tax}
\title{R's federal income tax}
\source{
General Social Survey https://gss.norc.org
}
\description{
tax
}
\details{
Question Do you consider the amount of federal income tax which you have to pay as too high, about right, or too low?
}
\section{Values}{

\itemize{
\item \code{1} too high
\item \code{2} about right
\item \code{3} too low
\item \code{4} r pays no income tax (vol.)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1976 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab A/C \tab partial \cr
   1989 \tab A/C \tab partial \cr
   1990 \tab A/C \tab partial \cr
   1991 \tab A/C \tab partial \cr
   1993 \tab A/C \tab partial \cr
   1994 \tab A/C \tab partial \cr
   1996 \tab A/C \tab partial \cr
   1998 \tab A/C \tab partial \cr
   2000 \tab A/C \tab partial \cr
   2002 \tab A/C \tab partial \cr
   2004 \tab A/C \tab partial \cr
   2006 \tab A/C \tab partial \cr
   2008 \tab A/C \tab partial \cr
   2010 \tab A/C \tab partial \cr
   2012 \tab A/C \tab partial \cr
   2014 \tab A/C \tab partial \cr
   2016 \tab A/C \tab partial \cr
   2018 \tab A/C \tab partial \cr
   2021 \tab A/C \tab partial \cr
   2022 \tab A/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/245/vshow}{GSS Data Explorer page for \code{tax}}.
}

\seealso{
Other Core: 
\code{\link{colath}},
\code{\link{colsoc}},
\code{\link{educop}},
\code{\link{eqwlth}},
\code{\link{eqwlthy}},
\code{\link{govcare}},
\code{\link{libath}},
\code{\link{spkath}},
\code{\link{spksoc}}

Other Split Ballots: 
\code{\link{eqwlth}},
\code{\link{eqwlthy}}
}
\concept{Core}
\concept{Split Ballots}
\concept{Taxes}
\keyword{variable}
