% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_447.R
\name{taxbylaw}
\alias{taxbylaw}
\title{Do authorities tax legally}
\source{
General Social Survey https://gss.norc.org
}
\description{
taxbylaw
}
\details{
Question (In general, how often do you think that the tax authorities in America do the following?)
Treat everyone in accordance with the law, regardless of their contacts or position in society?  (Would you say...)
}
\section{Values}{

\itemize{
\item \code{1} almost always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} almost never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6312/vshow}{GSS Data Explorer page for \code{taxbylaw}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab almost always \tab often \tab sometimes \tab almost never \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1477 \tab 385 \tab 268 \tab 464 \tab 212 \tab 47 \tab 14 \tab 2867 \cr
   Total \tab 1477 \tab 385 \tab 268 \tab 464 \tab 212 \tab 47 \tab 14 \tab 2867 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{brlawfl}},
\code{\link{brnotax}},
\code{\link{demworks}},
\code{\link{poleff17}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{runpower}},
\code{\link{taxpaid}},
\code{\link{taxspend}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
