% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_509.R
\name{taxcheat}
\alias{taxcheat}
\title{Wrong to cheat on taxes?}
\source{
General Social Survey https://gss.norc.org
}
\description{
taxcheat
}
\details{
Question Consider the situations listed below. Do you feel it is wrong or not wrong if. . .
A. A taxpayer does not report all of his income in order to pay less income taxes.
}
\section{Values}{

\itemize{
\item \code{1} not wrong
\item \code{2} a bit wrong
\item \code{3} wrong
\item \code{4} seriously wrong
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4637/vshow}{GSS Data Explorer page for \code{taxcheat}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{abdefct1}},
\code{\link{abdefctw}},
\code{\link{abpoor1}},
\code{\link{abpoorw}},
\code{\link{conbiz}},
\code{\link{concong}},
\code{\link{congovt}},
\code{\link{govcheat}},
\code{\link{homosex1}}

Other Federal Government: 
\code{\link{concong}},
\code{\link{congovt}},
\code{\link{govcheat}}
}
\concept{Federal Government}
\concept{ISSP Religion}
\keyword{variable}
