% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_473.R
\name{taxmid}
\alias{taxmid}
\title{Taxes on middle income people too high}
\source{
General Social Survey https://gss.norc.org
}
\description{
taxmid
}
\details{
Question Generally, how would you describe taxes in America today...
We mean all taxes together, including social security, income tax, sales tax, and all the rest.
B. Next, for those with middle incomes, are taxes . . .
}
\section{Values}{

\itemize{
\item \verb{[1]} much too high
\item \verb{[2]} too high
\item \verb{[3]} about right
\item \verb{[4]} too low
\item \verb{[5]} much too low
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1996 \cr
   D/-/- \tab 2006 \cr
   Full \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4303/vshow}{GSS Data Explorer page for \code{taxmid}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab much too high \tab too high \tab about right \tab too low \tab much too low \tab can't choose \tab no answer \tab Total \cr
   1987 \tab 255 \tab 251 \tab 777 \tab 408 \tab 35 \tab 4 \tab 66 \tab 23 \tab 1819 \cr
   1996 \tab 1572 \tab 208 \tab 615 \tab 394 \tab 27 \tab 2 \tab 54 \tab 32 \tab 2904 \cr
   2006 \tab 2992 \tab 206 \tab 635 \tab 574 \tab 46 \tab 11 \tab 46 \tab - \tab 4510 \cr
   2016 \tab 1477 \tab 150 \tab 580 \tab 557 \tab 60 \tab 6 \tab 22 \tab 15 \tab 2867 \cr
   Total \tab 6296 \tab 815 \tab 2607 \tab 1933 \tab 168 \tab 23 \tab 188 \tab 70 \tab 12100 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conclass}},
\code{\link{conjobs}},
\code{\link{consoc}},
\code{\link{conunion}},
\code{\link{conurban}},
\code{\link{conwlth}},
\code{\link{taxpoor}},
\code{\link{taxrich}},
\code{\link{taxshare}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
