% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_473.R
\name{taxpoor}
\alias{taxpoor}
\title{Taxes on low income people too high}
\source{
General Social Survey https://gss.norc.org
}
\description{
taxpoor
}
\details{
Question Generally, how would you describe taxes in America today...
We mean all taxes together, including social security, income tax, sales tax, and all the rest.
C. Lastly, for those with low incomes, are taxes . . .
}
\section{Values}{

\itemize{
\item \verb{[1]} much too high
\item \verb{[2]} too high
\item \verb{[3]} about right
\item \verb{[4]} too low
\item \verb{[5]} much too low
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1996 \cr
   D/-/- \tab 2006 \cr
   Full \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4304/vshow}{GSS Data Explorer page for \code{taxpoor}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab much too high \tab too high \tab about right \tab too low \tab much too low \tab can't choose \tab no answer \tab Total \cr
   1987 \tab 255 \tab 477 \tab 575 \tab 352 \tab 43 \tab 15 \tab 75 \tab 27 \tab 1819 \cr
   1996 \tab 1572 \tab 292 \tab 493 \tab 368 \tab 46 \tab 7 \tab 89 \tab 37 \tab 2904 \cr
   2006 \tab 2992 \tab 378 \tab 529 \tab 472 \tab 60 \tab 22 \tab 56 \tab 1 \tab 4510 \cr
   2016 \tab 1477 \tab 313 \tab 501 \tab 424 \tab 81 \tab 30 \tab 27 \tab 14 \tab 2867 \cr
   Total \tab 6296 \tab 1460 \tab 2098 \tab 1616 \tab 230 \tab 74 \tab 247 \tab 79 \tab 12100 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conclass}},
\code{\link{conjobs}},
\code{\link{consoc}},
\code{\link{conunion}},
\code{\link{conurban}},
\code{\link{conwlth}},
\code{\link{taxmid}},
\code{\link{taxrich}},
\code{\link{taxshare}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
