% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_473.R
\name{taxrich}
\alias{taxrich}
\title{Taxes on high income people too high}
\source{
General Social Survey https://gss.norc.org
}
\description{
taxrich
}
\details{
Question Generally, how would you describe taxes in America today...
We mean all taxes together, including social security, income tax, sales tax, and all the rest.
A. First, for those with high incomes, are taxes . . .
}
\section{Values}{

\itemize{
\item \verb{[1]} much too high
\item \verb{[2]} too high
\item \verb{[3]} about right
\item \verb{[4]} too low
\item \verb{[5]} much too low
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1996 \cr
   B/C/- \tab 2008, 2021 \cr
   D/-/- \tab 2006 \cr
   Full \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4302/vshow}{GSS Data Explorer page for \code{taxrich}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab much too high \tab too high \tab about right \tab too low \tab much too low \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1987 \tab 255 \tab 114 \tab 173 \tab 319 \tab 558 \tab 275 \tab 99 \tab 26 \tab - \tab 1819 \cr
   1996 \tab 1572 \tab 140 \tab 310 \tab 296 \tab 360 \tab 112 \tab 87 \tab 27 \tab - \tab 2904 \cr
   2006 \tab 2992 \tab 70 \tab 160 \tab 383 \tab 519 \tab 284 \tab 100 \tab 2 \tab - \tab 4510 \cr
   2008 \tab 982 \tab 170 \tab 214 \tab 245 \tab 265 \tab 104 \tab 39 \tab 4 \tab - \tab 2023 \cr
   2016 \tab 1477 \tab 104 \tab 210 \tab 326 \tab 482 \tab 212 \tab 40 \tab 16 \tab - \tab 2867 \cr
   2021 \tab 2029 \tab 98 \tab 240 \tab 329 \tab 614 \tab 423 \tab 143 \tab 144 \tab 12 \tab 4032 \cr
   Total \tab 9307 \tab 696 \tab 1307 \tab 1898 \tab 2798 \tab 1410 \tab 508 \tab 219 \tab 12 \tab 18155 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conclass}},
\code{\link{conjobs}},
\code{\link{consoc}},
\code{\link{conunion}},
\code{\link{conurban}},
\code{\link{conwlth}},
\code{\link{taxmid}},
\code{\link{taxpoor}},
\code{\link{taxshare}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
