% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_87.R
\name{taxserve}
\alias{taxserve}
\title{Favor conscription given taxes?}
\source{
General Social Survey https://gss.norc.org
}
\description{
taxserve
}
\details{
Question C. IF STRONGLY FAVOR OR PROBABLY FAVOR EITHER A OR B: and suppose that the costs of such a program made it necessary to increase your taxes by a small amount--for example, 5 percent. Would you strongly favor it, probably favor it, probably oppose it, or strongly oppose it?
}
\section{Values}{

\itemize{
\item \code{1} strongly favor
\item \code{2} probably favor
\item \code{3} probably oppose
\item \code{4} strongly oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/780/vshow}{GSS Data Explorer page for \code{taxserve}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly favor \tab probably favor \tab probably oppose \tab strongly oppose \tab don't know \tab no answer \tab Total \cr
   1982 \tab 484 \tab 262 \tab 517 \tab 299 \tab 216 \tab 38 \tab 44 \tab 1860 \cr
   1984 \tab 400 \tab 193 \tab 460 \tab 250 \tab 106 \tab 21 \tab 43 \tab 1473 \cr
   Total \tab 884 \tab 455 \tab 977 \tab 549 \tab 322 \tab 59 \tab 87 \tab 3333 \cr
}
}

\seealso{
Other Core: 
\code{\link{blnumok}},
\code{\link{blnumoky}},
\code{\link{fenumok}},
\code{\link{fightair}},
\code{\link{hinumok}},
\code{\link{hinumoky}},
\code{\link{milpay}},
\code{\link{milqual}},
\code{\link{milvolok}}
}
\concept{Core}
\keyword{variable}
