% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_473.R
\name{taxshare}
\alias{taxshare}
\title{Should rich pay bigger share of taxes?}
\source{
General Social Survey https://gss.norc.org
}
\description{
taxshare
}
\details{
Question Do you think that people with high incomes should pay a larger share of their income in taxes than those with low incomes, the same share, or a smaller share . . .
}
\section{Values}{

\itemize{
\item \code{1} much larger share
\item \code{2} larger
\item \code{3} the same share
\item \code{4} smaller
\item \code{5} much smaller share
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab -/-/- \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2008 \tab B/C/- \tab partial \cr
   2021 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4305/vshow}{GSS Data Explorer page for \code{taxshare}}.
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conclass}},
\code{\link{conjobs}},
\code{\link{consoc}},
\code{\link{conunion}},
\code{\link{conurban}},
\code{\link{conwlth}},
\code{\link{taxmid}},
\code{\link{taxpoor}},
\code{\link{taxrich}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
