% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_473.R
\name{taxshare}
\alias{taxshare}
\title{Should rich pay bigger share of taxes?}
\source{
General Social Survey https://gss.norc.org
}
\description{
taxshare
}
\details{
Question Do you think that people with high incomes should pay a larger share of their income in taxes than those with low incomes, the same share, or a smaller share . . .
}
\section{Values}{

\itemize{
\item \code{1} much larger share
\item \code{2} larger
\item \code{3} the same share
\item \code{4} smaller
\item \code{5} much smaller share
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
   B/C/- \tab 2008, 2021 \cr
   Full \tab 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4305/vshow}{GSS Data Explorer page for \code{taxshare}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab much larger share \tab larger \tab the same share \tab smaller \tab much smaller share \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   1987 \tab 255 \tab 347 \tab 648 \tab 428 \tab 24 \tab 7 \tab 88 \tab 22 \tab - \tab 1819 \cr
   2000 \tab 1545 \tab 260 \tab 502 \tab 384 \tab 16 \tab 11 \tab 49 \tab 50 \tab - \tab 2817 \cr
   2008 \tab 982 \tab 196 \tab 421 \tab 380 \tab 15 \tab 5 \tab 18 \tab 6 \tab - \tab 2023 \cr
   2021 \tab 2029 \tab 512 \tab 760 \tab 432 \tab 32 \tab 12 \tab 101 \tab 144 \tab 10 \tab 4032 \cr
   Total \tab 4811 \tab 1315 \tab 2331 \tab 1624 \tab 87 \tab 35 \tab 256 \tab 222 \tab 10 \tab 10691 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{conclass}},
\code{\link{conjobs}},
\code{\link{consoc}},
\code{\link{conunion}},
\code{\link{conurban}},
\code{\link{conwlth}},
\code{\link{taxmid}},
\code{\link{taxpoor}},
\code{\link{taxrich}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
