% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_447.R
\name{taxspend}
\alias{taxspend}
\title{Choice between reducing taxes or spending more}
\source{
General Social Survey https://gss.norc.org
}
\description{
taxspend
}
\details{
Question If the government had a choice between reducing taxes or spending more on social programs like health care, social security, and unemployment benefits, which do you think it should do? (We mean all taxes together, including social security, income tax, sales tax, and all the rest.) PLEASE CHECK ONE BOX ONLY.
}
\section{Values}{

\itemize{
\item \verb{[1]} reduce taxes
\item \verb{[2]} spend more on social programs
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4103/vshow}{GSS Data Explorer page for \code{taxspend}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab reduce taxes \tab spend more on social programs \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1572 \tab 408 \tab 610 \tab 292 \tab 22 \tab 2904 \cr
   Total \tab 1572 \tab 408 \tab 610 \tab 292 \tab 22 \tab 2904 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{brlawfl}},
\code{\link{brnotax}},
\code{\link{demworks}},
\code{\link{poleff17}},
\code{\link{runbanks}},
\code{\link{runhosp}},
\code{\link{runpower}},
\code{\link{taxbylaw}},
\code{\link{taxpaid}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
