% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_503.R
\name{techwork}
\alias{techwork}
\title{New techs make work more interesting}
\source{
General Social Survey https://gss.norc.org
}
\description{
techwork
}
\details{
Question Do you think that the introduction of new technologies in the United States over the next few years will make work…
}
\section{Values}{

\itemize{
\item \code{1} much more
\item \code{2} little more
\item \code{3} neither mre/less
\item \code{4} a little less
\item \code{5} much less
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4576/vshow}{GSS Data Explorer page for \code{techwork}}.
}

\seealso{
Other Work Orientation: 
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paytime}},
\code{\link{placewrk}},
\code{\link{sethours}},
\code{\link{techjobs}},
\code{\link{wantjob1}},
\code{\link{worknow}}

Other Job: 
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paytime}},
\code{\link{placewrk}},
\code{\link{sethours}},
\code{\link{techjobs}},
\code{\link{wantjob1}}

Other Technology: 
\code{\link{techjobs}}

Other Work: 
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paytime}},
\code{\link{placewrk}},
\code{\link{sethours}},
\code{\link{techjobs}},
\code{\link{wantjob1}},
\code{\link{worknow}}
}
\concept{Job}
\concept{Technology}
\concept{Work}
\concept{Work Orientation}
\keyword{variable}
