% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_69.R
\name{teenpill}
\alias{teenpill}
\title{Birth control information to teens}
\source{
General Social Survey https://gss.norc.org
}
\description{
teenpill
}
\details{
Question B. Do you think birth control information should be available to teenagers who want it, or not?
}
\section{Values}{

\itemize{
\item \code{1} available
\item \code{2} not available
\item \code{3} depends on age
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1974, 1975, 1977, 1982, 1983 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/624/vshow}{GSS Data Explorer page for \code{teenpill}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab available \tab not available \tab depends on age \tab don't know \tab no answer \tab Total \cr
   1974 \tab - \tab 1145 \tab 286 \tab 3 \tab 37 \tab 13 \tab 1484 \cr
   1975 \tab - \tab 1167 \tab 278 \tab - \tab 38 \tab 7 \tab 1490 \cr
   1977 \tab - \tab 1249 \tab 244 \tab - \tab 30 \tab 7 \tab 1530 \cr
   1982 \tab - \tab 1580 \tab 225 \tab - \tab 50 \tab 5 \tab 1860 \cr
   1983 \tab - \tab 1354 \tab 199 \tab - \tab 39 \tab 7 \tab 1599 \cr
   Total \tab 0 \tab 6495 \tab 1232 \tab 3 \tab 194 \tab 39 \tab 7963 \cr
}
}

\seealso{
Other Core: 
\code{\link{abcare}},
\code{\link{abfirm}},
\code{\link{chldidel}},
\code{\link{chldmore}},
\code{\link{chldnum}},
\code{\link{chldsoon}},
\code{\link{pill}},
\code{\link{pillok}},
\code{\link{sexeduc}}
}
\concept{Core}
\keyword{variable}
