% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_69.R
\name{teenpill}
\alias{teenpill}
\title{Birth control information to teens}
\source{
General Social Survey https://gss.norc.org
}
\description{
teenpill
}
\details{
Question B. Do you think birth control information should be available to teenagers who want it, or not?
}
\section{Values}{

\itemize{
\item \code{1} available
\item \code{2} not available
\item \code{3} depends on age
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1974 \tab - \tab full \cr
   1975 \tab - \tab full \cr
   1977 \tab - \tab full \cr
   1982 \tab - \tab full \cr
   1983 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/624/vshow}{GSS Data Explorer page for \code{teenpill}}.
}

\seealso{
Other Core: 
\code{\link{abcare}},
\code{\link{abfirm}},
\code{\link{chldidel}},
\code{\link{chldmore}},
\code{\link{chldnum}},
\code{\link{chldsoon}},
\code{\link{pill}},
\code{\link{pillok}},
\code{\link{sexeduc}}

Other Birth Control: 
\code{\link{pill}},
\code{\link{pillok}}
}
\concept{Birth Control}
\concept{Core}
\keyword{variable}
