% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_484.R
\name{teensex1}
\alias{teensex1}
\title{Morality of teens having sex}
\source{
General Social Survey https://gss.norc.org
}
\description{
teensex1
}
\details{
Question What if they are in their early teens, say under 16 years old, in that case is it . . .
}
\section{Values}{

\itemize{
\item \verb{[1]} always wrong
\item \verb{[2]} almost always wrong
\item \verb{[3]} wrong only sometimes
\item \verb{[4]} not wrong at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4415/vshow}{GSS Data Explorer page for \code{teensex1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab always wrong \tab almost always wrong \tab wrong only sometimes \tab not wrong at all \tab can't choose \tab no answer \tab Total \cr
   1994 \tab 1545 \tab 927 \tab 215 \tab 114 \tab 54 \tab 48 \tab 89 \tab 2992 \cr
   Total \tab 1545 \tab 927 \tab 215 \tab 114 \tab 54 \tab 48 \tab 89 \tab 2992 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{abchoose}},
\code{\link{caresick}},
\code{\link{chldcare}},
\code{\link{fambudgt}},
\code{\link{laundry}},
\code{\link{livnowed}},
\code{\link{mapaid}},
\code{\link{repairs}},
\code{\link{sexhar}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
