% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_523.R
\name{tempgen}
\alias{tempgen}
\title{Greenhouse effect danger to envir}
\source{
General Social Survey https://gss.norc.org
}
\description{
tempgen
}
\details{
Question A. In general, do you think that a rise in the world's temperature caused by the `greenhouse effect', is…
}
\section{Values}{

\itemize{
\item \verb{[1]} extremely dangerous
\item \verb{[2]} very dangerous
\item \verb{[3]} somewhat dangerous
\item \verb{[4]} not very dangerous
\item \verb{[5]} not dangerous at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/-/- \tab 2016, 2018 \cr
   A/B/- \tab 2021 \cr
   A/B/C \tab 1993, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4764/vshow}{GSS Data Explorer page for \code{tempgen}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab extremely dangerous \tab very dangerous \tab somewhat dangerous \tab not very dangerous \tab not dangerous at all \tab don't know \tab no answer \tab Total \cr
   1993 \tab 49 \tab 263 \tab 370 \tab 510 \tab 187 \tab 33 \tab 158 \tab 36 \tab 1606 \cr
   1994 \tab 1606 \tab 186 \tab 287 \tab 489 \tab 184 \tab 30 \tab 177 \tab 33 \tab 2992 \cr
   2000 \tab 1541 \tab 202 \tab 312 \tab 420 \tab 116 \tab 21 \tab 153 \tab 52 \tab 2817 \cr
   2016 \tab 1956 \tab 246 \tab 240 \tab 247 \tab 98 \tab 38 \tab 40 \tab 2 \tab 2867 \cr
   2018 \tab 1563 \tab 253 \tab 197 \tab 205 \tab 75 \tab 23 \tab 28 \tab 4 \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   Total \tab 10747 \tab 1150 \tab 1406 \tab 1871 \tab 660 \tab 145 \tab 556 \tab 127 \tab 16662 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{chemgen}},
\code{\link{indusgen}},
\code{\link{watergen}}

Other ISSP Environment: 
\code{\link{busdecid}},
\code{\link{chemfam}},
\code{\link{chemgen}},
\code{\link{indusfam}},
\code{\link{indusgen}},
\code{\link{pubdecid}},
\code{\link{tempfam}},
\code{\link{waterfam}},
\code{\link{watergen}}
}
\concept{ISSP Environment}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
