% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_528.R
\name{tempgen1}
\alias{tempgen1}
\title{Greenhouse effect danger to envir}
\source{
General Social Survey https://gss.norc.org
}
\description{
tempgen1
}
\details{
Question H. In general, do you think that a rise in the world’s temperature caused by climate change is . . .
}
\section{Values}{

\itemize{
\item \code{1} extremely dangerous
\item \code{2} very dangerous
\item \code{3} somewhat dangerous
\item \code{4} not very dangerous
\item \code{5} not dangerous at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/B \tab partial \cr
   2021 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4810/vshow}{GSS Data Explorer page for \code{tempgen1}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{busgrn}},
\code{\link{grneffme}},
\code{\link{h2oless}},
\code{\link{helpharm}},
\code{\link{nobuygrn}},
\code{\link{peopgrn}},
\code{\link{redcehme}}

Other Environment: 
\code{\link{busgrn}},
\code{\link{grneffme}},
\code{\link{h2oless}},
\code{\link{helpharm}},
\code{\link{nobuygrn}},
\code{\link{peopgrn}},
\code{\link{redcehme}}
}
\concept{Environment}
\concept{ISSP Environment}
\concept{Knowledge}
\concept{Science}
\keyword{variable}
