% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_250.R
\name{termvig}
\alias{termvig}
\title{One of the vignettes}
\source{
General Social Survey https://gss.norc.org
}
\description{
termvig
}
\details{
Question Now I'm going to describe a patient to you and then ask you some questions about that patient.
}
\section{Values}{

\itemize{
\item \code{1} vignette a
\item \code{2} vignette b
\item \code{3} vignette c
\item \code{4} vignette d
\item \code{5} vignette e
\item \code{6} vignette f
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2254/vshow}{GSS Data Explorer page for \code{termvig}}.
}

\seealso{
Other Mental Health: 
\code{\link{docsui}},
\code{\link{hospice}},
\code{\link{instrctn}},
\code{\link{notreat}},
\code{\link{painmed}},
\code{\link{rightref}},
\code{\link{trustcrt}},
\code{\link{trustdoc}},
\code{\link{trustfam}}
}
\concept{Mental Health}
\keyword{variable}
