% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_138.R
\name{thisjob3}
\alias{thisjob3}
\title{Heard job from my supervisor}
\source{
General Social Survey https://gss.norc.org
}
\description{
thisjob3
}
\details{
Question Which of the answers on this card describe how you heard about your present job with your employer (IF SELF-EMPLOYED:  current organization)? Please tell me all of the ways that you found out about this job.
C. From my supervisor.
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1294/vshow}{GSS Data Explorer page for \code{thisjob3}}.
}

\seealso{
Other Work Organizations: 
\code{\link{jobyears}},
\code{\link{promotng}},
\code{\link{samejob}},
\code{\link{thisjob1}},
\code{\link{thisjob2}},
\code{\link{thisjob4}},
\code{\link{thisjob5}},
\code{\link{thisjob6}},
\code{\link{thisjob7}}
}
\concept{Work Organizations}
\keyword{variable}
