% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_188.R
\name{thnkangr}
\alias{thnkangr}
\title{How often thought about this situation}
\source{
General Social Survey https://gss.norc.org
}
\description{
thnkangr
}
\details{
Question Even after they go away, feelings sometimes come back when we think about things. How often have you thought about this situation since it happened: very often, every once in a while or just once or twice, or never (until asked  about it just now)?
}
\section{Values}{

\itemize{
\item \code{1} very often
\item \code{2} every once in a while
\item \code{3} just once or twice
\item \code{4} never-not once until now
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1766/vshow}{GSS Data Explorer page for \code{thnkangr}}.
}

\seealso{
Other Emotions: 
\code{\link{angrlast}},
\code{\link{chnang1}},
\code{\link{chnang2}},
\code{\link{chnang3}},
\code{\link{chnang4}},
\code{\link{chnang5}},
\code{\link{howangry}},
\code{\link{madat22}},
\code{\link{madat23}}

Other Feelings: 
\code{\link{angrlast}},
\code{\link{chnang1}},
\code{\link{chnang2}},
\code{\link{chnang3}},
\code{\link{chnang4}},
\code{\link{chnang5}},
\code{\link{howangry}},
\code{\link{madat22}},
\code{\link{madat23}}
}
\concept{Emotions}
\concept{Feelings}
\keyword{variable}
