% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_487.R
\name{timehome}
\alias{timehome}
\title{Run out of time before getting things done at home}
\source{
General Social Survey https://gss.norc.org
}
\description{
timehome
}
\details{
Question To what extent do you agree or disagree?
A. There are so many things to do at home, I often run out of time before I get them all done.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4443/vshow}{GSS Data Explorer page for \code{timehome}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2002 \tab 1594 \tab 384 \tab 399 \tab 170 \tab 153 \tab 46 \tab 16 \tab 3 \tab 2765 \cr
   Total \tab 1594 \tab 384 \tab 399 \tab 170 \tab 153 \tab 46 \tab 16 \tab 3 \tab 2765 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{buythngs}},
\code{\link{famvswrk}},
\code{\link{jobvsfam}},
\code{\link{strsshme}},
\code{\link{strsswrk}},
\code{\link{timework}},
\code{\link{tiredhme}},
\code{\link{tiredwrk}},
\code{\link{weekend}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
