% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_291.R
\name{timekid1}
\alias{timekid1}
\title{Time r spends with child 1}
\source{
General Social Survey https://gss.norc.org
}
\description{
timekid1
}
\details{
Question Would you say that the amount of time you spend with (NAME OF CHILD) is too much, about the right amount, or too little?
A. First listed child?
}
\section{Values}{

\itemize{
\item \code{1} too much
\item \code{2} about the right amount
\item \code{3} too little
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2639/vshow}{GSS Data Explorer page for \code{timekid1}}.
}

\seealso{
Other Information Society: 
\code{\link{advsrch}},
\code{\link{browser3}},
\code{\link{ezines}},
\code{\link{mp3}},
\code{\link{newsgrps}},
\code{\link{prefsets}},
\code{\link{timekid2}},
\code{\link{timekid3}},
\code{\link{timekid4}}
}
\concept{Information Society}
\keyword{variable}
