% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_291.R
\name{timekid4}
\alias{timekid4}
\title{Time r spends with child 4}
\source{
General Social Survey https://gss.norc.org
}
\description{
timekid4
}
\details{
Question Would you say that the amount of time you spend with (NAME OF CHILD) is too much, about the right amount, or too little?
D. Fourth listed child?
}
\section{Values}{

\itemize{
\item \verb{[1]} too much
\item \verb{[2]} about the right amount
\item \verb{[3]} too little
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2642/vshow}{GSS Data Explorer page for \code{timekid4}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab too much \tab about the right amount \tab too little \tab no answer \tab Total \cr
   2000 \tab 2742 \tab 3 \tab 34 \tab 20 \tab 18 \tab 2817 \cr
   Total \tab 2742 \tab 3 \tab 34 \tab 20 \tab 18 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{advsrch}},
\code{\link{browser3}},
\code{\link{ezines}},
\code{\link{mp3}},
\code{\link{newsgrps}},
\code{\link{prefsets}},
\code{\link{timekid1}},
\code{\link{timekid2}},
\code{\link{timekid3}}
}
\concept{Information Society}
\keyword{variable}
