% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_507.R
\name{timeoff}
\alias{timeoff}
\title{How difficult r takes an hour off during working hours}
\source{
General Social Survey https://gss.norc.org
}
\description{
timeoff
}
\details{
Question How difficult would it be for you to take an hour or two off during working hours, to take care of personal or family matters?
}
\section{Values}{

\itemize{
\item \code{1} not difficult at all
\item \code{2} not too difficult
\item \code{3} somewhat difficult
\item \code{4} very difficult
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab D/- \tab partial \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4613/vshow}{GSS Data Explorer page for \code{timeoff}}.
}

\seealso{
Other Work Orientation: 
\code{\link{dailywrk}},
\code{\link{famintjb}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobeasy}},
\code{\link{newjob}},
\code{\link{nounemp1}},
\code{\link{nounemp2}},
\code{\link{replaceu}}
}
\concept{Work Orientation}
\keyword{variable}
