% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_507.R
\name{timeoff}
\alias{timeoff}
\title{How difficult r takes an hour off during working hours}
\source{
General Social Survey https://gss.norc.org
}
\description{
timeoff
}
\details{
Question How difficult would it be for you to take an hour or two off during working hours, to take care of personal or family matters?
}
\section{Values}{

\itemize{
\item \code{1} not difficult at all
\item \code{2} not too difficult
\item \code{3} somewhat difficult
\item \code{4} very difficult
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
   D/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4613/vshow}{GSS Data Explorer page for \code{timeoff}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab not difficult at all \tab not too difficult \tab somewhat difficult \tab very difficult \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3492 \tab 448 \tab 274 \tab 161 \tab 132 \tab 1 \tab 2 \tab 4510 \cr
   2016 \tab 1927 \tab 397 \tab 289 \tab 144 \tab 103 \tab 2 \tab 5 \tab 2867 \cr
   Total \tab 5419 \tab 845 \tab 563 \tab 305 \tab 235 \tab 3 \tab 7 \tab 7377 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{dailywrk}},
\code{\link{famintjb}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobeasy}},
\code{\link{newjob}},
\code{\link{nounemp1}},
\code{\link{nounemp2}},
\code{\link{replaceu}}
}
\concept{Work Orientation}
\keyword{variable}
