% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_491.R
\name{timerelx}
\alias{timerelx}
\title{Time r would spend relaxing}
\source{
General Social Survey https://gss.norc.org
}
\description{
timerelx
}
\details{
Question Suppose you could change the way you spend your time, spending more on some things and less on others. Which of these things on the following list would you like to spend more time on, which would you like to spend less time on, and which would you like to spend the same amount of time as now?
F. Time to relax?
}
\section{Values}{

\itemize{
\item \code{1} spend much more
\item \code{2} spend a bit more
\item \code{3} spend same
\item \code{4} spend a bit less
\item \code{5} spend much less
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4491/vshow}{GSS Data Explorer page for \code{timerelx}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab spend much more \tab spend a bit more \tab spend same \tab spend a bit less \tab spend much less \tab can't choose \tab no answer \tab iap \tab Total \cr
   1989 \tab - \tab 328 \tab 513 \tab 489 \tab 48 \tab 30 \tab 17 \tab 112 \tab - \tab 1537 \cr
   Total \tab 0 \tab 328 \tab 513 \tab 489 \tab 48 \tab 30 \tab 17 \tab 112 \tab 0 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{timefam}},
\code{\link{timefrnd}},
\code{\link{timehhwk}},
\code{\link{timeleis}},
\code{\link{timepdwk}}
}
\concept{Work Orientation}
\keyword{variable}
