% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_487.R
\name{timework}
\alias{timework}
\title{Run out of time before getting things done at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
timework
}
\details{
Question To what extent do you agree or disagree?
C. There are so many things to do at work, I often run out of time before I get them all done.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4445/vshow}{GSS Data Explorer page for \code{timework}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2002 \tab 1594 \tab 203 \tab 274 \tab 213 \tab 252 \tab 60 \tab 159 \tab 10 \tab 2765 \cr
   Total \tab 1594 \tab 203 \tab 274 \tab 213 \tab 252 \tab 60 \tab 159 \tab 10 \tab 2765 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{buythngs}},
\code{\link{famvswrk}},
\code{\link{jobvsfam}},
\code{\link{strsshme}},
\code{\link{strsswrk}},
\code{\link{timehome}},
\code{\link{tiredhme}},
\code{\link{tiredwrk}},
\code{\link{weekend}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
