% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_490.R
\name{tiredhm1}
\alias{tiredhm1}
\title{How often too tired to do housework}
\source{
General Social Survey https://gss.norc.org
}
\description{
tiredhm1
}
\details{
Question How often has each of the following happened to you during the past three months?
A. You have come home from work too tired to do the chores which need to be done.
}
\section{Values}{

\itemize{
\item \code{0} doesn't apply/no job
\item \code{1} several times a week
\item \code{2} several times a month
\item \code{3} once or twice
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/- \tab 2022 \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4475/vshow}{GSS Data Explorer page for \code{tiredhm1}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab doesn't apply/no job \tab several times a week \tab several times a month \tab once or twice \tab never \tab don't know \tab no answer \tab 5 \tab skipped on web \tab Total \cr
   2012 \tab 672 \tab 412 \tab 279 \tab 183 \tab 252 \tab 164 \tab 4 \tab 8 \tab - \tab - \tab 1974 \cr
   2022 \tab 2386 \tab - \tab 278 \tab 206 \tab 225 \tab 143 \tab 5 \tab 6 \tab 289 \tab 6 \tab 3544 \cr
   Total \tab 3058 \tab 412 \tab 557 \tab 389 \tab 477 \tab 307 \tab 9 \tab 14 \tab 289 \tab 6 \tab 5518 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{eldcost}},
\code{\link{famvswk1}},
\code{\link{hhclean1}},
\code{\link{jobvsfa1}},
\code{\link{tiredwk1}},
\code{\link{wkkidscl}},
\code{\link{wkkidscs}},
\code{\link{wkndact}},
\code{\link{wkyngscl}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
