% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_490.R
\name{tiredwk1}
\alias{tiredwk1}
\title{How often too tired from housework to do job well}
\source{
General Social Survey https://gss.norc.org
}
\description{
tiredwk1
}
\details{
Question How often has each of the following happened to you during the past three months?
C. You have arrived at work too tired to function well because of the household work you had done.
}
\section{Values}{

\itemize{
\item \code{0} doesn't apply/no job
\item \code{1} several times a week
\item \code{2} several times a month
\item \code{3} once or twice
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/- \tab 2022 \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4477/vshow}{GSS Data Explorer page for \code{tiredwk1}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab doesn't apply/no job \tab several times a week \tab several times a month \tab once or twice \tab never \tab don't know \tab no answer \tab 5 \tab skipped on web \tab Total \cr
   2012 \tab 672 \tab 452 \tab 34 \tab 43 \tab 153 \tab 609 \tab 3 \tab 8 \tab - \tab - \tab 1974 \cr
   2022 \tab 2386 \tab - \tab 43 \tab 84 \tab 180 \tab 497 \tab 8 \tab 4 \tab 331 \tab 11 \tab 3544 \cr
   Total \tab 3058 \tab 452 \tab 77 \tab 127 \tab 333 \tab 1106 \tab 11 \tab 12 \tab 331 \tab 11 \tab 5518 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{eldcost}},
\code{\link{famvswk1}},
\code{\link{hhclean1}},
\code{\link{jobvsfa1}},
\code{\link{tiredhm1}},
\code{\link{wkkidscl}},
\code{\link{wkkidscs}},
\code{\link{wkndact}},
\code{\link{wkyngscl}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
