% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_490.R
\name{tiredwk1}
\alias{tiredwk1}
\title{How often too tired from housework to do job well}
\source{
General Social Survey https://gss.norc.org
}
\description{
tiredwk1
}
\details{
Question How often has each of the following happened to you during the past three months?
C. You have arrived at work too tired to function well because of the household work you had done.
}
\section{Values}{

\itemize{
\item \code{0} doesn't apply/no job
\item \code{1} several times a week
\item \code{2} several times a month
\item \code{3} once or twice
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B \tab partial \cr
   2022 \tab A/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4477/vshow}{GSS Data Explorer page for \code{tiredwk1}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{eldcost}},
\code{\link{famvswk1}},
\code{\link{hhclean1}},
\code{\link{jobvsfa1}},
\code{\link{tiredhm1}},
\code{\link{wkkidscl}},
\code{\link{wkkidscs}},
\code{\link{wkndact}},
\code{\link{wkyngscl}}

Other Household: 
\code{\link{hhclean1}},
\code{\link{tiredhm1}}
}
\concept{Household}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
