% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_213.R
\name{tlkclrgy}
\alias{tlkclrgy}
\title{X should talk to religious leader}
\source{
General Social Survey https://gss.norc.org
}
\description{
tlkclrgy
}
\details{
Question Should NAME do any of the following:
B. Talk to a minister, priest, rabbi or other religious leader
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2012/vshow}{GSS Data Explorer page for \code{tlkclrgy}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 1174 \tab 169 \tab 70 \tab 31 \tab 2904 \cr
   2006 \tab 3073 \tab 1207 \tab 175 \tab 40 \tab 15 \tab 4510 \cr
   2018 \tab 1175 \tab 970 \tab 144 \tab 29 \tab 30 \tab 2348 \cr
   Total \tab 5708 \tab 3351 \tab 488 \tab 139 \tab 76 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{dofirst}},
\code{\link{meddoc}},
\code{\link{mentldoc}},
\code{\link{mentlhos}},
\code{\link{mentloth}},
\code{\link{otcmed}},
\code{\link{rxmed}},
\code{\link{selfhelp}},
\code{\link{sphealer}}
}
\concept{Mental Health}
\keyword{variable}
