% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_212.R
\name{tlkfam}
\alias{tlkfam}
\title{X should talk to family and friends}
\source{
General Social Survey https://gss.norc.org
}
\description{
tlkfam
}
\details{
Question Should NAME do any of the following:
A. Talk things over with family and friends
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2011/vshow}{GSS Data Explorer page for \code{tlkfam}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 1355 \tab 54 \tab 13 \tab 22 \tab 2904 \cr
   2006 \tab 3073 \tab 1348 \tab 53 \tab 21 \tab 15 \tab 4510 \cr
   2018 \tab 1175 \tab 1107 \tab 25 \tab 13 \tab 28 \tab 2348 \cr
   Total \tab 5708 \tab 3810 \tab 132 \tab 47 \tab 65 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{hurtoth}},
\code{\link{hurtself}},
\code{\link{imprvtrt}},
\code{\link{vigfrnd}},
\code{\link{viggrp}},
\code{\link{vigmar}},
\code{\link{vignei}},
\code{\link{vigsoc}},
\code{\link{vigwork}}
}
\concept{Mental Health}
\keyword{variable}
