% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_427.R
\name{toldsxor}
\alias{toldsxor}
\title{At what age you first told someone that you were gay/les/bi}
\source{
General Social Survey https://gss.norc.org
}
\description{
toldsxor
}
\details{
Question At about what age were you when you first told someone that you were (gay(IF MALE)/lesbian(IF FEMALE)/bisexual)?
}
\section{Values}{

\itemize{
\item \code{1} under age 14
\item \code{2} age 14-18
\item \code{3} age 18-25
\item \code{4} age 26-35
\item \code{5} age 36-45
\item \code{6} age 46-55
\item \code{7} age 56-65
\item \code{8} age 66 or older
\item \code{9} i have never told anyone
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3909/vshow}{GSS Data Explorer page for \code{toldsxor}}.
}

\seealso{
Other Sexual Orientation: 
\code{\link{adoptkid}},
\code{\link{attractd}},
\code{\link{dwell5}},
\code{\link{emphlth1}},
\code{\link{emphlth2}},
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evkid}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{hlthcovr}},
\code{\link{hlthtype}},
\code{\link{kidlived}},
\code{\link{kidnow1}},
\code{\link{kidnow2}},
\code{\link{kidresp}},
\code{\link{losejob5}},
\code{\link{marunion}},
\code{\link{mykids}},
\code{\link{negjob5}},
\code{\link{othkids}},
\code{\link{relgendr}},
\code{\link{sexsex18}},
\code{\link{sexunion}},
\code{\link{sppart}},
\code{\link{toldsmsx}},
\code{\link{toldwork}}
}
\concept{Sexual Orientation}
\keyword{variable}
