% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_428.R
\name{toldwork}
\alias{toldwork}
\title{How many of r's coworkers know r is gay/les/bi}
\source{
General Social Survey https://gss.norc.org
}
\description{
toldwork
}
\details{
Question Some people are very open about their sexual orientation while others are not. About how many  of your coworkers know that you (are gay(IF MALE)/are lesbian(IF FEMALE)/are bisexual/have sex with men(IF MALE)/have sex with women(IF FEMALE))?
}
\section{Values}{

\itemize{
\item \code{1} all of my coworkers
\item \code{2} more than half
\item \code{3} about half
\item \code{4} less than half
\item \code{5} none of them
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008, 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3919/vshow}{GSS Data Explorer page for \code{toldwork}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab all of my coworkers \tab more than half \tab about half \tab less than half \tab none of them \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1928 \tab 12 \tab 4 \tab 4 \tab 15 \tab 22 \tab 2 \tab 36 \tab 2023 \cr
   2012 \tab 1902 \tab 10 \tab 5 \tab 7 \tab 10 \tab 34 \tab 1 \tab 5 \tab 1974 \cr
   Total \tab 3830 \tab 22 \tab 9 \tab 11 \tab 25 \tab 56 \tab 3 \tab 41 \tab 3997 \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{dwell5}},
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{kidnow2}},
\code{\link{losejob5}},
\code{\link{negjob5}}
}
\concept{Sexual Orientation}
\keyword{variable}
