% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_428.R
\name{toldwork}
\alias{toldwork}
\title{How many of r's coworkers know r is gay/les/bi}
\source{
General Social Survey https://gss.norc.org
}
\description{
toldwork
}
\details{
Question Some people are very open about their sexual orientation while others are not. About how many  of your coworkers know that you (are gay(IF MALE)/are lesbian(IF FEMALE)/are bisexual/have sex with men(IF MALE)/have sex with women(IF FEMALE))?
}
\section{Values}{

\itemize{
\item \code{1} all of my coworkers
\item \code{2} more than half
\item \code{3} about half
\item \code{4} less than half
\item \code{5} none of them
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3919/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab about half \tab all of my coworkers \tab don't know \tab less than half \tab more than half \tab no answer \tab none of them \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 1928 \tab 4 \tab 12 \tab 2 \tab 15 \tab 4 \tab 36 \tab 22 \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1902 \tab 7 \tab 10 \tab 1 \tab 10 \tab 5 \tab 5 \tab 34 \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 68679 \tab 11 \tab 22 \tab 3 \tab 25 \tab 9 \tab 41 \tab 56 \tab 6853 \tab 75699 \cr
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
}
}

\seealso{
Other Sexual Orientation: 
\code{\link{dwell5}},
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{kidnow2}},
\code{\link{losejob5}},
\code{\link{negjob5}}

Other Gays And Lesbians: 
\code{\link{dwell5}},
\code{\link{evdwell}},
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{evnegjob}},
\code{\link{harjob5}},
\code{\link{losejob5}},
\code{\link{negjob5}}

Other Job: 
\code{\link{evharjb}},
\code{\link{evlosejb}},
\code{\link{harjob5}},
\code{\link{negjob5}}

Other Work: 
\code{\link{evharjb}},
\code{\link{harjob5}},
\code{\link{negjob5}}
}
\concept{Gays And Lesbians}
\concept{Job}
\concept{Sexual Orientation}
\concept{Work}
\keyword{variable}
