% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_370.R
\name{tomatoes}
\alias{tomatoes}
\title{Regular tomatoes no genes}
\source{
General Social Survey https://gss.norc.org
}
\description{
tomatoes
}
\details{
Question Now, I would like to ask you a few short questions like those you might see on a television game show. for each statement that I read, please tell me if it is true or false. If you don't know or aren't sure, just tell me so,  and we will skip to the next question. Remember true, false, or don't know.
L. Ordinary tomatoes do not contain genes, while genetically modified tomatoes Do.
}
\section{Values}{

\itemize{
\item \verb{[1]} true
\item \verb{[2]} false
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/C \tab 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3345/vshow}{GSS Data Explorer page for \code{tomatoes}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab true \tab false \tab can't choose \tab no answer \tab Total \cr
   2010 \tab 1103 \tab 181 \tab 445 \tab 311 \tab 4 \tab 2044 \cr
   Total \tab 1103 \tab 181 \tab 445 \tab 311 \tab 4 \tab 2044 \cr
}
}

\seealso{
Other Science Knowledge & Attitudes: 
\code{\link{coldeg1}},
\code{\link{condrift}},
\code{\link{earthsun}},
\code{\link{evolved}},
\code{\link{evolved1}},
\code{\link{evolved2}},
\code{\link{intrhome}},
\code{\link{majorcol}},
\code{\link{solarrev}}
}
\concept{Science Knowledge & Attitudes}
\keyword{variable}
