% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_521.R
\name{toodifme}
\alias{toodifme}
\title{Too diff to do anything about envir}
\source{
General Social Survey https://gss.norc.org
}
\description{
toodifme
}
\details{
Question How much do you agree or disagree with each of these statements?
A. It is just too difficult for someone like me to do much about the environment.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2010 \tab A/B/- \tab partial \cr
   2021 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4739/vshow}{GSS Data Explorer page for \code{toodifme}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{grnsol}},
\code{\link{grntest1}},
\code{\link{grntest2}},
\code{\link{ihlpgrn}},
\code{\link{scitest1}},
\code{\link{scitest2}},
\code{\link{scitest3}},
\code{\link{scitest4}},
\code{\link{scitest5}}

Other Environment: 
\code{\link{grnsol}},
\code{\link{grntest1}},
\code{\link{grntest2}},
\code{\link{ihlpgrn}}
}
\concept{Environment}
\concept{ISSP Environment}
\keyword{variable}
