% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_527.R
\name{topprob1}
\alias{topprob1}
\title{Most important issue for america}
\source{
General Social Survey https://gss.norc.org
}
\description{
topprob1
}
\details{
Question A1. Which of these issues is the most important for America today?
}
\section{Values}{

\itemize{
\item \verb{[1]} health care
\item \verb{[2]} education
\item \verb{[3]} crime
\item \verb{[4]} the environment
\item \verb{[5]} immigration
\item \verb{[6]} the economy
\item \verb{[7]} terrorism
\item \verb{[8]} poverty
\item \verb{[9]} none of these
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2010, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4798/vshow}{GSS Data Explorer page for \code{topprob1}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab health care \tab education \tab crime \tab the environment \tab immigration \tab the economy \tab terrorism \tab poverty \tab none of these \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2010 \tab 614 \tab 309 \tab 258 \tab 19 \tab 58 \tab 53 \tab 492 \tab 107 \tab 37 \tab 13 \tab 78 \tab 6 \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 551 \tab 216 \tab 53 \tab 224 \tab 68 \tab 363 \tab 50 \tab 99 \tab 66 \tab 197 \tab 140 \tab 2 \tab 4032 \cr
   Total \tab 2617 \tab 860 \tab 474 \tab 72 \tab 282 \tab 121 \tab 855 \tab 157 \tab 136 \tab 79 \tab 275 \tab 146 \tab 2 \tab 6076 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{enprbfam}},
\code{\link{enprbus}},
\code{\link{fair5}},
\code{\link{futenrgy}},
\code{\link{grncon}},
\code{\link{knowsol}},
\code{\link{knwcause}},
\code{\link{topprob2}},
\code{\link{trust5}}
}
\concept{ISSP Environment}
\keyword{variable}
