% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_527.R
\name{topprob2}
\alias{topprob2}
\title{Next important issue}
\source{
General Social Survey https://gss.norc.org
}
\description{
topprob2
}
\details{
Question A2. Which is the next most important?
}
\section{Values}{

\itemize{
\item \verb{[1]} health care
\item \verb{[2]} education
\item \verb{[3]} crime
\item \verb{[4]} the environment
\item \verb{[5]} immigration
\item \verb{[6]} the economy
\item \verb{[7]} terrorism
\item \verb{[8]} poverty
\item \verb{[9]} none of these
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2010, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4799/vshow}{GSS Data Explorer page for \code{topprob2}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab health care \tab education \tab crime \tab the environment \tab immigration \tab the economy \tab terrorism \tab poverty \tab none of these \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2010 \tab 614 \tab 327 \tab 325 \tab 73 \tab 120 \tab 76 \tab 203 \tab 152 \tab 71 \tab 6 \tab 69 \tab 8 \tab - \tab 2044 \cr
   2021 \tab 2009 \tab 463 \tab 332 \tab 139 \tab 201 \tab 88 \tab 312 \tab 52 \tab 159 \tab 61 \tab 42 \tab 143 \tab 31 \tab 4032 \cr
   Total \tab 2623 \tab 790 \tab 657 \tab 212 \tab 321 \tab 164 \tab 515 \tab 204 \tab 230 \tab 67 \tab 111 \tab 151 \tab 31 \tab 6076 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{enprbfam}},
\code{\link{enprbus}},
\code{\link{fair5}},
\code{\link{futenrgy}},
\code{\link{grncon}},
\code{\link{knowsol}},
\code{\link{knwcause}},
\code{\link{topprob1}},
\code{\link{trust5}}
}
\concept{ISSP Environment}
\keyword{variable}
