% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_527.R
\name{topprob2}
\alias{topprob2}
\title{Next important issue}
\source{
General Social Survey https://gss.norc.org
}
\description{
topprob2
}
\details{
Question A2. Which is the next most important?
}
\section{Values}{

\itemize{
\item \code{1} health care
\item \code{2} education
\item \code{3} crime
\item \code{4} the environment
\item \code{5} immigration
\item \code{6} the economy
\item \code{7} terrorism
\item \code{8} poverty
\item \code{9} none of these
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/B \tab partial \cr
   2021 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4799/vshow}{GSS Data Explorer page for \code{topprob2}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{enprbfam}},
\code{\link{enprbus}},
\code{\link{fair5}},
\code{\link{futenrgy}},
\code{\link{grncon}},
\code{\link{knowsol}},
\code{\link{knwcause}},
\code{\link{topprob1}},
\code{\link{trust5}}
}
\concept{ISSP Environment}
\keyword{variable}
