% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_140.R
\name{totsup}
\alias{totsup}
\title{How many people responsible to you}
\source{
General Social Survey https://gss.norc.org
}
\description{
totsup
}
\details{
Question Counting all of these levels, how may people are responsible to you both directly and indirectly?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1317/vshow}{GSS Data Explorer page for \code{totsup}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 100 \tab 11 \tab 12 \tab 1200 \tab 125 \tab 13 \tab 137 \tab 14 \tab 15 \tab 16 \tab 18 \tab 19 \tab 2 \tab 20 \tab 200 \tab 21 \tab 22 \tab 2211 \tab 23 \tab 25 \tab 26 \tab 27 \tab 28 \tab 3 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 39 \tab 4 \tab 40 \tab 400 \tab 45 \tab 47 \tab 5 \tab 50 \tab 55 \tab 58 \tab 6 \tab 60 \tab 65 \tab 7 \tab 70 \tab 8 \tab 9 \tab 90 \tab 96 \tab don't know \tab no answer \tab Total \cr
   1991 \tab 1176 \tab 51 \tab 15 \tab 3 \tab 1 \tab 12 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 10 \tab 2 \tab 2 \tab 1 \tab 36 \tab 13 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 34 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 34 \tab 2 \tab 1 \tab 1 \tab 1 \tab 23 \tab 3 \tab 1 \tab 1 \tab 11 \tab 3 \tab 2 \tab 5 \tab 2 \tab 13 \tab 3 \tab 1 \tab 1 \tab 5 \tab 17 \tab 1517 \cr
   Total \tab 1176 \tab 51 \tab 15 \tab 3 \tab 1 \tab 12 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 10 \tab 2 \tab 2 \tab 1 \tab 36 \tab 13 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 34 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4 \tab 1 \tab 34 \tab 2 \tab 1 \tab 1 \tab 1 \tab 23 \tab 3 \tab 1 \tab 1 \tab 11 \tab 3 \tab 2 \tab 5 \tab 2 \tab 13 \tab 3 \tab 1 \tab 1 \tab 5 \tab 17 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{automatn}},
\code{\link{idecide}},
\code{\link{levels}()},
\code{\link{monitred}},
\code{\link{offsup}},
\code{\link{supduty1}},
\code{\link{supothrs}},
\code{\link{yousup}}
}
\concept{Work Organizations}
\keyword{variable}
