% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_207.R
\name{totwork}
\alias{totwork}
\title{How much money contributed in work-related org}
\source{
General Social Survey https://gss.norc.org
}
\description{
totwork
}
\details{
Question Approximately how much money or the cash equivalent of property have you contributed in each of the fields listed above in the past twelve months? (REPEAT FOR EACH ITEM MENTIONED IN GIVHLTH-GIVOTH) RECORD ABOVE UNDER TOTHLTH-VALOTH. (ESTIMATED DOLLAR VALUE CONTRIBUTED INCLUDING BOTH CASH CONTRIBUTIONS AND THE CASH-VALUE OF PROPERTY.)
I. Work-related organizations
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1953/vshow}{GSS Data Explorer page for \code{totwork}}.
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{totart}},
\code{\link{totfound}},
\code{\link{totinfrm}},
\code{\link{totintl}},
\code{\link{tototh}},
\code{\link{totpol}},
\code{\link{totyouth}},
\code{\link{valeduc}},
\code{\link{valhlth}}

Other Voluntary Associations: 
\code{\link{totart}},
\code{\link{totfound}},
\code{\link{totinfrm}},
\code{\link{totintl}},
\code{\link{tototh}},
\code{\link{totpol}},
\code{\link{totyouth}},
\code{\link{valeduc}},
\code{\link{valhlth}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\concept{Voluntary Associations}
\keyword{variable}
