% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_177.R
\name{tradmod}
\alias{tradmod}
\title{Man alone-both responsible for providing income}
\source{
General Social Survey https://gss.norc.org
}
\description{
tradmod
}
\details{
Question Next, we have some questions about the kind of relationship you would like with a spouse or partner. It doesn't matter whether you are now married or living with someone. For each pair of items on the card, which type of relationship would you prefer? In pair A, would you prefer the type of relationship described in statement 1 or statement 2? How about in pair B? Pair C?
A.
}
\section{Values}{

\itemize{
\item \verb{[1]} man alone responsible
\item \verb{[2]} both responsible
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1657/vshow}{GSS Data Explorer page for \code{tradmod}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab man alone responsible \tab both responsible \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1444 \tab 431 \tab 1020 \tab 4 \tab 5 \tab 2904 \cr
   Total \tab 1444 \tab 431 \tab 1020 \tab 4 \tab 5 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{emoteoth}},
\code{\link{fejobaff}},
\code{\link{hapboys}},
\code{\link{hapgirls}},
\code{\link{rhmewrk}},
\code{\link{sharesep}},
\code{\link{sphmewrk}},
\code{\link{willwed1}},
\code{\link{willwed2}}
}
\concept{Gender}
\keyword{variable}
