% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_309.R
\name{trainops}
\alias{trainops}
\title{R have the training opportunities}
\source{
General Social Survey https://gss.norc.org
}
\description{
trainops
}
\details{
Question Now I''m going to read you another list of statements about your main job. For each, please tell me if the statement is very true, somewhat true, not too true, or not at all true with respect to the work you do.
O. I have the training opportunities I need to perform my job safely and competently
}
\section{Values}{

\itemize{
\item \code{1} very true
\item \code{2} somewhat true
\item \code{3} not too true
\item \code{4} not at all true
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2810/vshow}{GSS Data Explorer page for \code{trainops}}.
}

\seealso{
Other Quality of Working Life: 
\code{\link{cowrkhlp}},
\code{\link{fairearn}},
\code{\link{handmove}},
\code{\link{hvylift}},
\code{\link{manvsemp}},
\code{\link{suphelp}},
\code{\link{wkbonus}},
\code{\link{wkpraise}},
\code{\link{wrktime}}

Other Job: 
\code{\link{cowrkhlp}},
\code{\link{handmove}},
\code{\link{hvylift}},
\code{\link{manvsemp}},
\code{\link{suphelp}},
\code{\link{wkbonus}},
\code{\link{wkpraise}},
\code{\link{wrktime}}
}
\concept{Job}
\concept{Quality of Working Life}
\keyword{variable}
