% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_84.R
\name{trarel5}
\alias{trarel5}
\title{Trauma of relatives, last 5 yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
trarel5
}
\details{
Question B. Number of traumatic events (divorces, unemployment, hospitalizations/disabilities) happening to respondent''s relatives during last five years.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1978 \tab - \tab full \cr
   1980 \tab - \tab full \cr
   1983 \tab - \tab full \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/756/vshow}{GSS Data Explorer page for \code{trarel5}}.
}

\seealso{
Other Core: 
\code{\link{defspdfg}},
\code{\link{defspdr}},
\code{\link{sibdeath}},
\code{\link{spdeath}},
\code{\link{trarel1}},
\code{\link{tratot1}},
\code{\link{tratot5}},
\code{\link{trauma1}},
\code{\link{trauma5}}

Other Trauma: 
\code{\link{trarel1}},
\code{\link{tratot1}},
\code{\link{tratot5}},
\code{\link{trauma1}},
\code{\link{trauma5}}
}
\concept{Core}
\concept{Trauma}
\keyword{variable}
