% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_338.R
\name{travelus}
\alias{travelus}
\title{Transportation r took when travelling to us}
\source{
General Social Survey https://gss.norc.org
}
\description{
travelus
}
\details{
Question Thinking about the most recent time you traveled to the United States from another country—please tell me which box shows the kind of transportation you took: A, B, or C.
}
\section{Values}{

\itemize{
\item \verb{[1]} box a
\item \verb{[2]} box b
\item \verb{[3]} box c
\item \verb{[4]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3026/vshow}{GSS Data Explorer page for \code{travelus}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab box a \tab box b \tab other \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2539 \tab 21 \tab 225 \tab 7 \tab 12 \tab 8 \tab 2812 \cr
   Total \tab 2539 \tab 21 \tab 225 \tab 7 \tab 12 \tab 8 \tab 2812 \cr
}
}

\seealso{
Other Immigration: 
\code{\link{bornhome}},
\code{\link{immstats}},
\code{\link{natborn}},
\code{\link{robject}},
\code{\link{rundrstd}}
}
\concept{Immigration}
\keyword{variable}
