% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_678.R
\name{trcong}
\alias{trcong}
\title{Trust in congress, 10 point scale}
\source{
General Social Survey https://gss.norc.org
}
\description{
trcong
}
\details{
Question On a scale of 0 to 10, how much do you personally trust each of the following institutions? 0 means you do not trust an institution at all, and 10 means you trust it completely….The U.S. Congress.
}
\section{Values}{

\itemize{
\item \code{0} no trust at all
\item \code{10} complete trust
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8176/vshow}{GSS Data Explorer page for \code{trcong}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{agetech}},
\code{\link{aidrive}},
\code{\link{aimed}},
\code{\link{aiworry}},
\code{\link{classtech}},
\code{\link{dataprot}},
\code{\link{eductech}},
\code{\link{emonitor1}},
\code{\link{ftruse}},
\code{\link{gendtech}},
\code{\link{harmgood1}},
\code{\link{infodeal}},
\code{\link{infoprofit}},
\code{\link{intaccprxy}},
\code{\link{intapps}},
\code{\link{intbhlf}},
\code{\link{intcnfrm}},
\code{\link{intcomm}},
\code{\link{intfncl}},
\code{\link{intgame}},
\code{\link{inthome}},
\code{\link{inthrss}},
\code{\link{intliteracy}},
\code{\link{intlnly}},
\code{\link{intmeet}},
\code{\link{intnews}},
\code{\link{intoppo}},
\code{\link{intpolview}},
\code{\link{intprfssnl}},
\code{\link{intpub}},
\code{\link{intrnetuse}},
\code{\link{intrust}},
\code{\link{intscam}},
\code{\link{intsearch}},
\code{\link{intshare}},
\code{\link{intskill}},
\code{\link{intstream}},
\code{\link{intsurf}},
\code{\link{inttrig}},
\code{\link{intviews}},
\code{\link{modelobby}},
\code{\link{modeorg}},
\code{\link{modeorgprot}},
\code{\link{modepet}},
\code{\link{modeprot}},
\code{\link{nextgen1}},
\code{\link{nouseint1}},
\code{\link{nouseint10}},
\code{\link{nouseint11}},
\code{\link{nouseint12}},
\code{\link{nouseint2}},
\code{\link{nouseint3}},
\code{\link{nouseint4}},
\code{\link{nouseint5}},
\code{\link{nouseint6}},
\code{\link{nouseint7}},
\code{\link{nouseint8}},
\code{\link{nouseint9}},
\code{\link{papernews}},
\code{\link{polnewsfrom}},
\code{\link{prxyuse}},
\code{\link{radionews}},
\code{\link{smnews}},
\code{\link{techesy}},
\code{\link{trppl}},
\code{\link{tvnews1}},
\code{\link{unplgrsn}},
\code{\link{unplug}},
\code{\link{urbantech}},
\code{\link{webnews}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
