% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_678.R
\name{trcong}
\alias{trcong}
\title{Trust in congress, 10 point scale}
\source{
General Social Survey https://gss.norc.org
}
\description{
trcong
}
\details{
Question On a scale of 0 to 10, how much do you personally trust each of the following institutions? 0 means you do not trust an institution at all, and 10 means you trust it completely….The U.S. Congress.
}
\section{Values}{

\itemize{
\item \code{0} no trust at all
\item \code{10} complete trust
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8176/vshow}{GSS Data Explorer page for \code{trcong}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{agetech}},
\code{\link{classtech}},
\code{\link{dataprot}},
\code{\link{eductech}},
\code{\link{emonitor1}},
\code{\link{gendtech}},
\code{\link{infodeal}},
\code{\link{infoprofit}},
\code{\link{urbantech}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
