% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_296.R
\name{treat1}
\alias{treat1}
\title{R seek treat moderate activities}
\source{
General Social Survey https://gss.norc.org
}
\description{
treat1
}
\details{
Question Next, I’m going to read you some descriptions of health problems people sometimes experience. Some people might go to a doctor, others might not.
A. The first description is: not being able to do moderate activities, such as moving a table, pushing a vacuum cleaner, bowling or playing golf? How likely would you be to seek medical treatment–that is, see a doctor, a nurse or other health professional–for this situation?  Would you definitely go probably go, probably not go, or definitely not go?
}
\section{Values}{

\itemize{
\item \code{1} definitely go
\item \code{2} probably go
\item \code{3} probably not go
\item \code{4} definitely not go
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2688/vshow}{GSS Data Explorer page for \code{treat1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely go \tab probably go \tab probably not go \tab definitely not go \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 526 \tab 371 \tab 197 \tab 109 \tab 6 \tab 210 \tab 2817 \cr
   Total \tab 1398 \tab 526 \tab 371 \tab 197 \tab 109 \tab 6 \tab 210 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{downblue}},
\code{\link{energy}},
\code{\link{pain}},
\code{\link{peaceful}},
\code{\link{socacts}},
\code{\link{treat2}},
\code{\link{treat3}},
\code{\link{treat4}},
\code{\link{treat5}}
}
\concept{Health Status}
\keyword{variable}
