% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_297.R
\name{treat7}
\alias{treat7}
\title{R seek treat pain interfere with work}
\source{
General Social Survey https://gss.norc.org
}
\description{
treat7
}
\details{
Question Next, I’m going to read you some descriptions of health problems people sometimes experience. Some people might go to a doctor, others might not.
G. Having pain interfere with your normal work, including both work outside the home and housework?
}
\section{Values}{

\itemize{
\item \code{1} definitely go
\item \code{2} probably go
\item \code{3} probably not go
\item \code{4} definitely not go
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2694/vshow}{GSS Data Explorer page for \code{treat7}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely go \tab probably go \tab probably not go \tab definitely not go \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 563 \tab 418 \tab 143 \tab 69 \tab 7 \tab 219 \tab 2817 \cr
   Total \tab 1398 \tab 563 \tab 418 \tab 143 \tab 69 \tab 7 \tab 219 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{treat10}},
\code{\link{treat11}},
\code{\link{treat6}},
\code{\link{treat8}},
\code{\link{treat9}},
\code{\link{wait1}},
\code{\link{wait2}},
\code{\link{wait3}},
\code{\link{wait4}}
}
\concept{Health Status}
\keyword{variable}
