% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_628.R
\name{trmedia}
\alias{trmedia}
\title{How much do you trust the news media?}
\source{
General Social Survey https://gss.norc.org
}
\description{
trmedia
}
\details{
Question How much do you trust the news media?
}
\section{Values}{

\itemize{
\item \code{0} 0 - no trust
\item \code{10} 10 - complete trust
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7483/vshow}{GSS Data Explorer page for \code{trmedia}}.
}

\keyword{variable}
