% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_677.R
\name{trppl}
\alias{trppl}
\title{Trust in others, 10 point scale}
\source{
General Social Survey https://gss.norc.org
}
\description{
trppl
}
\details{
Question Generally  speaking, would you say that most people can be trusted or that you can’t be too careful in dealing with people; Where would you place yourself on this scale, where 0 means you can’t be too careful, and 10 means most people can be trusted.
}
\section{Values}{

\itemize{
\item \code{0} you can't be too careful
\item \code{10} most people can be trusted
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8174/vshow}{GSS Data Explorer page for \code{trppl}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{modelobby}},
\code{\link{modeorg}},
\code{\link{modeorgprot}},
\code{\link{modepet}},
\code{\link{modeprot}},
\code{\link{papernews}},
\code{\link{radionews}},
\code{\link{tvnews1}},
\code{\link{webnews}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
