% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_151.R
\name{trstprof}
\alias{trstprof}
\title{Trust profs judgments in assigning reading}
\source{
General Social Survey https://gss.norc.org
}
\description{
trstprof
}
\details{
Question I'm going to read several statements. After I read each one, please tell me whether you strongly agree, agree, disagree, or strongly disagree with it. For example, ...
B. I trust the judgment of the teachers and professors who decide what high school and college students should be reading.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} disagree
\item \verb{[4]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1423/vshow}{GSS Data Explorer page for \code{trstprof}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 69 \tab 886 \tab 482 \tab 75 \tab 81 \tab 13 \tab 1606 \cr
   1998 \tab 1387 \tab 65 \tab 772 \tab 408 \tab 112 \tab 74 \tab 14 \tab 2832 \cr
   Total \tab 1387 \tab 134 \tab 1658 \tab 890 \tab 187 \tab 155 \tab 27 \tab 4438 \cr
}
}

\seealso{
Other Culture: 
\code{\link{classics}},
\code{\link{english}},
\code{\link{grtbooks}},
\code{\link{judgeart}},
\code{\link{modpaint}},
\code{\link{pclit}},
\code{\link{tvnews}},
\code{\link{tvpbs}},
\code{\link{tvshows}}
}
\concept{Culture}
\keyword{variable}
