% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_405.R
\name{trtattnd}
\alias{trtattnd}
\title{Proportion of regular church attendee r trust}
\source{
General Social Survey https://gss.norc.org
}
\description{
trtattnd
}
\details{
Question C. How many do you think attend religious services regularly? (Would you say…)
}
\section{Values}{

\itemize{
\item \code{1} almost all
\item \code{2} most
\item \code{3} about half
\item \code{4} a few
\item \code{5} none
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3698/vshow}{GSS Data Explorer page for \code{trtattnd}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab almost all \tab most \tab about half \tab a few \tab none \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3821 \tab 79 \tab 111 \tab 183 \tab 156 \tab 45 \tab 99 \tab 16 \tab 4510 \cr
   Total \tab 3821 \tab 79 \tab 111 \tab 183 \tab 156 \tab 45 \tab 99 \tab 16 \tab 4510 \cr
}
}

\seealso{
Other Number Known: 
\code{\link{trtcohab}},
\code{\link{trtcon}},
\code{\link{trtdems}},
\code{\link{trtgay}},
\code{\link{trtgoatt}},
\code{\link{trtlib}},
\code{\link{trtnoatt}},
\code{\link{trtreps}},
\code{\link{trtwhite}}
}
\concept{Number Known}
\keyword{variable}
