% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_405.R
\name{trtdems}
\alias{trtdems}
\title{Proportion of democrat r trust}
\source{
General Social Survey https://gss.norc.org
}
\description{
trtdems
}
\details{
Question B. How many do you think are Democrats? (Would you say…)
}
\section{Values}{

\itemize{
\item \code{1} almost all
\item \code{2} most
\item \code{3} about half
\item \code{4} a few
\item \code{5} none
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3697/vshow}{GSS Data Explorer page for \code{trtdems}}.
}

\seealso{
Other Number Known: 
\code{\link{trtattnd}},
\code{\link{trtcohab}},
\code{\link{trtcon}},
\code{\link{trtgay}},
\code{\link{trtgoatt}},
\code{\link{trtlib}},
\code{\link{trtnoatt}},
\code{\link{trtreps}},
\code{\link{trtwhite}}
}
\concept{Number Known}
\keyword{variable}
