% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_50.R
\name{trust}
\alias{trust}
\title{Can people be trusted}
\source{
General Social Survey https://gss.norc.org
}
\description{
trust
}
\details{
Question A. Generally speaking, would you say that most people can be trusted or that you can't be too careful in dealing with people?
}
\section{Values}{

\itemize{
\item \code{1} most people can be trusted
\item \code{2} can't be too careful
\item \code{3} depends
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/- \tab full \cr
   1973 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1976 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab B/C/- \tab partial \cr
   1989 \tab B/C/- \tab partial \cr
   1990 \tab B/C/- \tab partial \cr
   1991 \tab B/C/- \tab partial \cr
   1993 \tab B/C/- \tab partial \cr
   1994 \tab B/C/- \tab partial \cr
   1996 \tab B/C/- \tab partial \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab B/C/- \tab partial \cr
   2002 \tab B/C/- \tab partial \cr
   2004 \tab B/C/- \tab partial \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab B/C/- \tab partial \cr
   2010 \tab B/C/- \tab partial \cr
   2012 \tab B/C/- \tab partial \cr
   2014 \tab B/C/- \tab partial \cr
   2016 \tab B/C/- \tab partial \cr
   2018 \tab B/C/- \tab partial \cr
   2021 \tab B/C/- \tab partial \cr
   2022 \tab B/C/- \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/441/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllll}{
   year \tab can't be too careful \tab depends \tab don't know \tab most people can be trusted \tab no answer \tab iap \tab not available in this year \tab Total \cr
   1972 \tab 803 \tab 56 \tab 15 \tab 738 \tab 1 \tab - \tab - \tab 1613 \cr
   1973 \tab 765 \tab 44 \tab 3 \tab 690 \tab 2 \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab - \tab - \tab - \tab - \tab 1484 \tab - \tab 1484 \cr
   1975 \tab 830 \tab 64 \tab 7 \tab 584 \tab 5 \tab - \tab - \tab 1490 \cr
   1976 \tab 775 \tab 56 \tab 2 \tab 664 \tab 2 \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab - \tab - \tab - \tab - \tab 1530 \tab - \tab 1530 \cr
   1978 \tab 866 \tab 66 \tab 4 \tab 596 \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 741 \tab 55 \tab 5 \tab 665 \tab 2 \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab - \tab - \tab - \tab - \tab 1860 \tab - \tab 1860 \cr
   1983 \tab 469 \tab 35 \tab 2 \tab 296 \tab 2 \tab 795 \tab - \tab 1599 \cr
   1984 \tab 724 \tab 37 \tab 4 \tab 700 \tab 8 \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab - \tab - \tab - \tab - \tab 1534 \tab - \tab 1534 \cr
   1986 \tab 876 \tab 39 \tab 4 \tab 551 \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1045 \tab 66 \tab 2 \tab 697 \tab 9 \tab - \tab - \tab 1819 \cr
   1988 \tab 562 \tab 43 \tab 5 \tab 385 \tab 2 \tab 484 \tab - \tab 1481 \cr
   1989 \tab 569 \tab 39 \tab 3 \tab 422 \tab 2 \tab 502 \tab - \tab 1537 \cr
   1990 \tab 517 \tab 39 \tab - \tab 342 \tab 1 \tab 473 \tab - \tab 1372 \cr
   1991 \tab 573 \tab 50 \tab 2 \tab 390 \tab 2 \tab 500 \tab - \tab 1517 \cr
   1993 \tab 630 \tab 45 \tab 4 \tab 377 \tab 1 \tab 549 \tab - \tab 1606 \cr
   1994 \tab 1229 \tab 85 \tab 7 \tab 681 \tab 9 \tab 981 \tab - \tab 2992 \cr
   1996 \tab 1166 \tab 99 \tab 6 \tab 650 \tab 4 \tab 979 \tab - \tab 2904 \cr
   1998 \tab 1326 \tab 137 \tab 7 \tab 883 \tab 5 \tab 474 \tab - \tab 2832 \cr
   2000 \tab 1093 \tab 124 \tab 9 \tab 662 \tab 8 \tab 921 \tab - \tab 2817 \cr
   2002 \tab 528 \tab 72 \tab 2 \tab 310 \tab - \tab 1853 \tab - \tab 2765 \cr
   2004 \tab 507 \tab 54 \tab 1 \tab 314 \tab - \tab 1936 \tab - \tab 2812 \cr
   2006 \tab 1521 \tab 174 \tab - \tab 814 \tab 1 \tab 2000 \tab - \tab 4510 \cr
   2008 \tab 859 \tab 56 \tab - \tab 450 \tab - \tab 658 \tab - \tab 2023 \cr
   2010 \tab 839 \tab 75 \tab 3 \tab 458 \tab 2 \tab 667 \tab - \tab 2044 \cr
   2012 \tab 847 \tab 53 \tab 7 \tab 431 \tab - \tab 636 \tab - \tab 1974 \cr
   2014 \tab 1076 \tab 88 \tab 3 \tab 519 \tab 1 \tab 851 \tab - \tab 2538 \cr
   2016 \tab 1244 \tab 79 \tab 5 \tab 623 \tab 5 \tab 911 \tab - \tab 2867 \cr
   2018 \tab 995 \tab 69 \tab 6 \tab 492 \tab 1 \tab 785 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 804 \tab 107 \tab 2 \tab 306 \tab - \tab 2325 \tab - \tab 3544 \cr
   2024 \tab 604 \tab 108 \tab 7 \tab 236 \tab 1 \tab 2353 \tab - \tab 3309 \cr
   Total \tab 25383 \tab 2114 \tab 127 \tab 15926 \tab 76 \tab 28041 \tab 4032 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{fair}},
\code{\link{hapcohab}},
\code{\link{health}},
\code{\link{helpful}},
\code{\link{life}},
\code{\link{trusty}}

Other People: 
\code{\link{fair}},
\code{\link{helpful}},
\code{\link{trusty}}

Other Split Ballots: 
\code{\link{satcity}},
\code{\link{sathobby}},
\code{\link{trusty}}
}
\concept{Core}
\concept{People}
\concept{Split Ballots}
\keyword{variable}
