% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_50.R
\name{trust1}
\alias{trust1}
\title{Can people be trusted (see codebook)}
\source{
General Social Survey https://gss.norc.org
}
\description{
trust1
}
\details{
Question C. Generally speaking, would you say that most people can be trusted or that you need to be very careful in dealing with people?
}
\section{Values}{

\itemize{
\item \verb{[1]} can trust
\item \verb{[2]} need to be very careful
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6644/vshow}{GSS Data Explorer page for \code{trust1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab can trust \tab need to be very careful \tab don't know \tab no answer \tab Total \cr
   2006 \tab 3085 \tab 465 \tab 937 \tab 18 \tab 5 \tab 4510 \cr
   Total \tab 3085 \tab 465 \tab 937 \tab 18 \tab 5 \tab 4510 \cr
}
}

\keyword{variable}
