% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_527.R
\name{trust5}
\alias{trust5}
\title{People can or cannot be trusted}
\source{
General Social Survey https://gss.norc.org
}
\description{
trust5
}
\details{
Question B1. Generally speaking, would you say that most people can be trusted, or that you can’t be too careful in dealing with people? please tell me what you think, where 1 means you can’t be too careful and 5 means most people can be trusted.
}
\section{Values}{

\itemize{
\item \code{1} you can't be too careful
\item \code{5} most people can be trusted
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2010, 2021 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4800/vshow}{GSS Data Explorer page for \code{trust5}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab you can't be too careful \tab 2 \tab 3 \tab 4 \tab most people can be trusted \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2010 \tab 614 \tab 388 \tab 187 \tab 380 \tab 286 \tab 162 \tab 23 \tab 4 \tab - \tab 2044 \cr
   2021 \tab 2003 \tab 355 \tab 286 \tab 644 \tab 373 \tab 170 \tab 41 \tab 154 \tab 6 \tab 4032 \cr
   Total \tab 2617 \tab 743 \tab 473 \tab 1024 \tab 659 \tab 332 \tab 64 \tab 158 \tab 6 \tab 6076 \cr
}
}

\seealso{
Other ISSP Environment: 
\code{\link{enprbfam}},
\code{\link{enprbus}},
\code{\link{fair5}},
\code{\link{futenrgy}},
\code{\link{grncon}},
\code{\link{knowsol}},
\code{\link{knwcause}},
\code{\link{topprob1}},
\code{\link{topprob2}}
}
\concept{ISSP Environment}
\keyword{variable}
