% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_527.R
\name{trust5}
\alias{trust5}
\title{People can or cannot be trusted}
\source{
General Social Survey https://gss.norc.org
}
\description{
trust5
}
\details{
Question B1. Generally speaking, would you say that most people can be trusted, or that you can’t be too careful in dealing with people? please tell me what you think, where 1 means you can’t be too careful and 5 means most people can be trusted.
}
\section{Values}{

\itemize{
\item \code{1} you can't be too careful
\item \code{5} most people can be trusted
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/B \tab partial \cr
   2021 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4800/vshow}{GSS Data Explorer page for \code{trust5}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{enprbfam}},
\code{\link{enprbus}},
\code{\link{fair5}},
\code{\link{futenrgy}},
\code{\link{grncon}},
\code{\link{knowsol}},
\code{\link{knwcause}},
\code{\link{topprob1}},
\code{\link{topprob2}}

Other People: 
\code{\link{fair5}}
}
\concept{ISSP Environment}
\concept{People}
\keyword{variable}
