% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_250.R
\name{trustdoc}
\alias{trustdoc}
\title{Would trust my doctor}
\source{
General Social Survey https://gss.norc.org
}
\description{
trustdoc
}
\details{
Question People sometimes are incapable of making decisions about their care and medical treatment at the end of life. If you were incapable, how much trust would you put in the following people to do what was best for you?
READ RESPONSE CATEGORIES
B. I would trust my doctor.
}
\section{Values}{

\itemize{
\item \code{1} completely
\item \code{2} a great deal
\item \code{3} somewhat
\item \code{4} only a little
\item \code{5} not at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2252/vshow}{GSS Data Explorer page for \code{trustdoc}}.
}

\seealso{
Other Mental Health: 
\code{\link{docsui}},
\code{\link{hospice}},
\code{\link{instrctn}},
\code{\link{notreat}},
\code{\link{painmed}},
\code{\link{rightref}},
\code{\link{termvig}},
\code{\link{trustcrt}},
\code{\link{trustfam}}
}
\concept{Mental Health}
\keyword{variable}
