% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_241.R
\name{trusting}
\alias{trusting}
\title{Most people can be trusted}
\source{
General Social Survey https://gss.norc.org
}
\description{
trusting
}
\details{
Question Do you think of yourself as a trusting person?  Are you...
}
\section{Values}{

\itemize{
\item \code{1} very trusting
\item \code{2} somewht trusting
\item \code{3} somwht dstrustng
\item \code{4} very distrusting
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2166/vshow}{GSS Data Explorer page for \code{trusting}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very trusting \tab somewht trusting \tab somwht dstrustng \tab very distrusting \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1445 \tab 737 \tab 540 \tab 85 \tab 21 \tab 3 \tab 1 \tab 2832 \cr
   Total \tab 1445 \tab 737 \tab 540 \tab 85 \tab 21 \tab 3 \tab 1 \tab 2832 \cr
}
}

\seealso{
Other Jobs: 
\code{\link{daily}},
\code{\link{hourly}},
\code{\link{monthly}},
\code{\link{weekly}},
\code{\link{yearly}}
}
\concept{Jobs}
\keyword{variable}
