% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_514.R
\name{trustsci}
\alias{trustsci}
\title{We trust too much in science}
\source{
General Social Survey https://gss.norc.org
}
\description{
trustsci
}
\details{
Question We trust too much in science and not enough in religious faith.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2018 \cr
   A/B/C \tab 1998 \cr
   B/C/- \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4687/vshow}{GSS Data Explorer page for \code{trustsci}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1548 \tab 106 \tab 270 \tab 329 \tab 347 \tab 155 \tab 42 \tab 35 \tab 2832 \cr
   2008 \tab 658 \tab 100 \tab 321 \tab 343 \tab 406 \tab 162 \tab 26 \tab 7 \tab 2023 \cr
   2018 \tab 1173 \tab 127 \tab 295 \tab 357 \tab 221 \tab 145 \tab 18 \tab 12 \tab 2348 \cr
   Total \tab 3379 \tab 333 \tab 886 \tab 1029 \tab 974 \tab 462 \tab 86 \tab 54 \tab 7203 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{antirel}},
\code{\link{befair}},
\code{\link{cantrust}},
\code{\link{godright}},
\code{\link{perright}},
\code{\link{religcon}},
\code{\link{religinf}},
\code{\link{religint}},
\code{\link{socright}}
}
\concept{ISSP Religion}
\keyword{variable}
