% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_638.R
\name{trustv}
\alias{trustv}
\title{Can people be trusted (with volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
trustv
}
\details{
Question Generally speaking, would you say that most people can be trusted or that you can't
be too careful in dealing with people?
}
\section{Values}{

\itemize{
\item \code{1} most people can be trusted
\item \code{2} can't be too careful
\item \code{3} depends (phone mode only: volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7593/vshow}{GSS Data Explorer page for \code{trustv}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab can trust \tab can't be too careful \tab depends (phone mode only: volunteered) \tab skipped on web \tab Total \cr
   2021 \tab 2673 \tab 295 \tab 557 \tab 505 \tab 2 \tab 4032 \cr
   2022 \tab 2974 \tab 98 \tab 257 \tab 214 \tab 1 \tab 3544 \cr
   Total \tab 5647 \tab 393 \tab 814 \tab 719 \tab 3 \tab 7576 \cr
}
}

\keyword{variable}
