% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_310.R
\name{trynewjb}
\alias{trynewjb}
\title{How likely r make effort for new job next year}
\source{
General Social Survey https://gss.norc.org
}
\description{
trynewjb
}
\details{
Question Taking everything into consideration, how likely is it you will make a genuine effort to find a new job with another employer within the next year?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2820/vshow}{GSS Data Explorer page for \code{trynewjb}}.
}

\seealso{
Other Quality of Working Life: 
\code{\link{health1}},
\code{\link{jobfind1}},
\code{\link{laidoff}},
\code{\link{rincblls}},
\code{\link{wkageism}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wkracism}},
\code{\link{wksexism}}

Other Workplace Conflict: 
\code{\link{health1}},
\code{\link{wkageism}},
\code{\link{wkracism}}
}
\concept{Quality of Working Life}
\concept{Workplace Conflict}
\keyword{variable}
