% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_77.R
\name{tvhours}
\alias{tvhours}
\title{Hours per day watching tv}
\source{
General Social Survey https://gss.norc.org
}
\description{
tvhours
}
\details{
Question On the average day, about how many hours do you personally watch television?
}
\section{Values}{

\itemize{
\item \code{0} 0 hours
\item \code{24} 24 hours
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/691/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 hours \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab 8 \tab 9 \tab no answer \tab 20 \tab 18 \tab 24 hours \tab don't know \tab 21 \tab 22 \tab 23 \tab skipped on web \tab 19 \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 57 \tab 256 \tab 15 \tab 1 \tab 10 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 395 \tab 291 \tab 217 \tab 116 \tab 76 \tab 19 \tab 23 \tab 1 \tab 7 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab - \tab 61 \tab 320 \tab 7 \tab 1 \tab 8 \tab 1 \tab - \tab 5 \tab 3 \tab - \tab 385 \tab 311 \tab 205 \tab 98 \tab 72 \tab 16 \tab 25 \tab 6 \tab 5 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 91 \tab 316 \tab 12 \tab - \tab 10 \tab 1 \tab - \tab 1 \tab - \tab - \tab 418 \tab 287 \tab 194 \tab 99 \tab 55 \tab 8 \tab 30 \tab 3 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab - \tab 111 \tab 261 \tab 15 \tab 2 \tab 5 \tab - \tab 2 \tab 2 \tab 2 \tab - \tab 351 \tab 276 \tab 192 \tab 127 \tab 65 \tab 15 \tab 24 \tab 2 \tab 14 \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 78 \tab 359 \tab 19 \tab 5 \tab 24 \tab 3 \tab 4 \tab 2 \tab 3 \tab - \tab 441 \tab 356 \tab 269 \tab 120 \tab 87 \tab 29 \tab 43 \tab 8 \tab 6 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 90 \tab 296 \tab 14 \tab - \tab 10 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab 401 \tab 326 \tab 212 \tab 102 \tab 89 \tab 9 \tab 36 \tab 3 \tab 3 \tab - \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab - \tab 78 \tab 269 \tab 13 \tab - \tab 5 \tab 2 \tab 3 \tab 2 \tab - \tab - \tab 396 \tab 311 \tab 209 \tab 121 \tab 65 \tab 11 \tab 33 \tab 1 \tab 10 \tab 4 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 65 \tab 261 \tab 22 \tab 3 \tab 11 \tab 1 \tab 2 \tab 1 \tab 3 \tab - \tab 389 \tab 298 \tab 193 \tab 100 \tab 61 \tab 18 \tab 33 \tab 2 \tab 4 \tab 1 \tab 2 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 493 \tab 32 \tab 184 \tab 10 \tab - \tab 7 \tab 2 \tab 1 \tab 2 \tab 2 \tab - \tab 243 \tab 188 \tab 139 \tab 77 \tab 56 \tab 7 \tab 23 \tab 3 \tab 7 \tab 3 \tab - \tab - \tab 2 \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 531 \tab 33 \tab 184 \tab 9 \tab 2 \tab 7 \tab 2 \tab 1 \tab - \tab - \tab - \tab 301 \tab 169 \tab 143 \tab 70 \tab 40 \tab 11 \tab 23 \tab - \tab 7 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 444 \tab 29 \tab 200 \tab 10 \tab - \tab 5 \tab - \tab - \tab 1 \tab - \tab 1 \tab 261 \tab 172 \tab 115 \tab 67 \tab 36 \tab 12 \tab 14 \tab 2 \tab 2 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 493 \tab 38 \tab 181 \tab 17 \tab - \tab 5 \tab - \tab 2 \tab - \tab 2 \tab - \tab 256 \tab 216 \tab 138 \tab 74 \tab 49 \tab 6 \tab 24 \tab 4 \tab 10 \tab 1 \tab - \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab - \tab 61 \tab 327 \tab 15 \tab - \tab 8 \tab - \tab - \tab - \tab 2 \tab - \tab 449 \tab 309 \tab 192 \tab 107 \tab 75 \tab 10 \tab 33 \tab 2 \tab 12 \tab 2 \tab - \tab 1 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 74 \tab 425 \tab 11 \tab 1 \tab 11 \tab - \tab 4 \tab 2 \tab - \tab - \tab 534 \tab 378 \tab 279 \tab 112 \tab 85 \tab 15 \tab 29 \tab 4 \tab 12 \tab - \tab - \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 944 \tab 85 \tab 384 \tab 22 \tab 5 \tab 12 \tab 2 \tab 2 \tab 3 \tab 1 \tab - \tab 564 \tab 343 \tab 261 \tab 123 \tab 80 \tab 19 \tab 31 \tab 3 \tab 10 \tab 1 \tab 3 \tab 1 \tab 3 \tab - \tab 2 \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 463 \tab 119 \tab 499 \tab 17 \tab 3 \tab 6 \tab 1 \tab 1 \tab 7 \tab - \tab - \tab 642 \tab 399 \tab 318 \tab 157 \tab 89 \tab 18 \tab 47 \tab 7 \tab 26 \tab 4 \tab 2 \tab - \tab 6 \tab 1 \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 940 \tab 107 \tab 380 \tab 24 \tab 4 \tab 21 \tab 1 \tab 2 \tab 7 \tab - \tab - \tab 510 \tab 310 \tab 233 \tab 95 \tab 64 \tab 18 \tab 47 \tab - \tab 45 \tab 3 \tab - \tab 2 \tab 3 \tab 1 \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 32 \tab 200 \tab 7 \tab 2 \tab 6 \tab - \tab - \tab - \tab 1 \tab - \tab 233 \tab 170 \tab 105 \tab 76 \tab 39 \tab 6 \tab 23 \tab 2 \tab - \tab 1 \tab - \tab 1 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 57 \tab 228 \tab 9 \tab 1 \tab 9 \tab 2 \tab 2 \tab 2 \tab 2 \tab - \tab 231 \tab 138 \tab 112 \tab 46 \tab 20 \tab 9 \tab 26 \tab 2 \tab - \tab 3 \tab - \tab - \tab 7 \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 79 \tab 422 \tab 23 \tab - \tab 14 \tab 1 \tab 7 \tab 2 \tab - \tab - \tab 577 \tab 337 \tab 226 \tab 136 \tab 99 \tab 23 \tab 34 \tab 4 \tab 2 \tab - \tab 2 \tab 1 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 694 \tab 87 \tab 273 \tab 18 \tab 1 \tab 9 \tab - \tab 2 \tab 2 \tab 2 \tab 1 \tab 354 \tab 215 \tab 184 \tab 63 \tab 59 \tab 10 \tab 35 \tab 3 \tab - \tab 1 \tab 1 \tab 4 \tab 5 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 614 \tab 103 \tab 277 \tab 7 \tab 1 \tab 12 \tab 3 \tab 5 \tab 1 \tab 1 \tab - \tab 373 \tab 258 \tab 150 \tab 105 \tab 64 \tab 23 \tab 29 \tab 4 \tab 2 \tab 2 \tab 2 \tab 4 \tab 2 \tab 1 \tab - \tab 1 \tab - \tab - \tab 2044 \cr
   2012 \tab 672 \tab 90 \tab 255 \tab 17 \tab - \tab 11 \tab 2 \tab 4 \tab 3 \tab 1 \tab - \tab 325 \tab 238 \tab 171 \tab 61 \tab 58 \tab 19 \tab 31 \tab 3 \tab 1 \tab 2 \tab 1 \tab 5 \tab 3 \tab - \tab 1 \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 863 \tab 120 \tab 310 \tab 17 \tab - \tab 14 \tab - \tab 2 \tab - \tab 3 \tab 1 \tab 437 \tab 293 \tab 227 \tab 113 \tab 75 \tab 11 \tab 37 \tab 1 \tab - \tab 2 \tab 1 \tab 5 \tab 6 \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 980 \tab 162 \tab 370 \tab 23 \tab 1 \tab 25 \tab 1 \tab 1 \tab 4 \tab 2 \tab - \tab 486 \tab 297 \tab 221 \tab 113 \tab 92 \tab 22 \tab 46 \tab 6 \tab - \tab 5 \tab 1 \tab 5 \tab 4 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 789 \tab 145 \tab 349 \tab 23 \tab - \tab 14 \tab - \tab 2 \tab 4 \tab 3 \tab 1 \tab 376 \tab 240 \tab 169 \tab 100 \tab 63 \tab 13 \tab 38 \tab 6 \tab 1 \tab 4 \tab 1 \tab 4 \tab 3 \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 1329 \tab 215 \tab 390 \tab 54 \tab 1 \tab 24 \tab - \tab 8 \tab 9 \tab 3 \tab 1 \tab 602 \tab 463 \tab 390 \tab 215 \tab 149 \tab 23 \tab 97 \tab 12 \tab - \tab 17 \tab 1 \tab 7 \tab - \tab 1 \tab - \tab 1 \tab 20 \tab - \tab 4032 \cr
   2022 \tab 1168 \tab 241 \tab 427 \tab 57 \tab 1 \tab 37 \tab 2 \tab 1 \tab 13 \tab - \tab 1 \tab 523 \tab 361 \tab 282 \tab 145 \tab 125 \tab 26 \tab 74 \tab 8 \tab 3 \tab 12 \tab 3 \tab 10 \tab 12 \tab - \tab 2 \tab - \tab 9 \tab 1 \tab 3544 \cr
   2024 \tab 1126 \tab 198 \tab 389 \tab 40 \tab - \tab 28 \tab 3 \tab 7 \tab 5 \tab 2 \tab - \tab 496 \tab 343 \tab 284 \tab 143 \tab 97 \tab 23 \tab 65 \tab 2 \tab 4 \tab 10 \tab 3 \tab 12 \tab 10 \tab 1 \tab 1 \tab - \tab 17 \tab - \tab 3309 \cr
   Total \tab 29231 \tab 2738 \tab 8992 \tab 547 \tab 35 \tab 368 \tab 31 \tab 69 \tab 84 \tab 40 \tab 7 \tab 11949 \tab 8293 \tab 6030 \tab 3081 \tab 2084 \tab 449 \tab 1053 \tab 104 \tab 196 \tab 85 \tab 28 \tab 65 \tab 77 \tab 6 \tab 8 \tab 2 \tab 46 \tab 1 \tab 75699 \cr
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1975 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1988 \tab A/B/- \tab partial \cr
   1989 \tab A/B/- \tab partial \cr
   1990 \tab A/B/- \tab partial \cr
   1991 \tab A/B/- \tab partial \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/- \tab partial \cr
   1996 \tab A/B/- \tab partial \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/- \tab partial \cr
   2002 \tab A/B/- \tab partial \cr
   2004 \tab A/B/- \tab partial \cr
   2006 \tab A/B/- \tab partial \cr
   2008 \tab A/B/- \tab partial \cr
   2010 \tab A/B/- \tab partial \cr
   2012 \tab A/B/- \tab partial \cr
   2014 \tab A/B/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
   2021 \tab A/B/- \tab partial \cr
   2022 \tab A/B/- \tab partial \cr
   2024 \tab A/B/- \tab partial \cr
}
}

\seealso{
Other Core: 
\code{\link{arrest}},
\code{\link{convictd}},
\code{\link{hunt}},
\code{\link{hunt1}},
\code{\link{huntothr}},
\code{\link{lockedup}},
\code{\link{news}()},
\code{\link{rowngun}},
\code{\link{ticket}}
}
\concept{Core}
\concept{Television Viewing}
\keyword{variable}
