% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_116.R
\name{tvrelig}
\alias{tvrelig}
\title{Hours per week r watches religious tv}
\source{
General Social Survey https://gss.norc.org
}
\description{
tvrelig
}
\details{
Question About how much time per week, in hours and minutes, do you normally spend watching religious shows on television?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1075/vshow}{GSS Data Explorer page for \code{tvrelig}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 0.1 \tab 0.2 \tab 0.3 \tab 0.4 \tab 0.5 \tab 0.8 \tab 1 \tab 1.1 \tab 1.4 \tab 1.5 \tab 1.8 \tab 2 \tab 2.5 \tab 3 \tab 3.5 \tab 4 \tab 4.5 \tab 5 \tab 6 \tab 7 \tab 8 \tab 8.5 \tab 0.6 \tab 0.7 \tab 1.2 \tab 1.3 \tab 2.2 \tab 2.8 \tab 4.1 \tab 9 \tab Total \cr
   1988 \tab - \tab 1051 \tab 5 \tab 2 \tab 15 \tab 2 \tab 60 \tab 5 \tab 184 \tab 1 \tab 3 \tab 15 \tab 1 \tab 68 \tab 3 \tab 24 \tab 1 \tab 13 \tab 1 \tab 8 \tab 6 \tab 5 \tab 7 \tab 1 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1998 \tab 1387 \tab 1013 \tab 4 \tab 11 \tab 7 \tab - \tab 60 \tab - \tab 182 \tab - \tab 1 \tab 16 \tab 1 \tab 75 \tab 9 \tab 23 \tab 1 \tab 10 \tab - \tab 8 \tab 2 \tab 5 \tab 1 \tab - \tab 1 \tab 6 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2832 \cr
   Total \tab 1387 \tab 2064 \tab 9 \tab 13 \tab 22 \tab 2 \tab 120 \tab 5 \tab 366 \tab 1 \tab 4 \tab 31 \tab 2 \tab 143 \tab 12 \tab 47 \tab 2 \tab 23 \tab 1 \tab 16 \tab 8 \tab 10 \tab 8 \tab 1 \tab 1 \tab 6 \tab 4 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 4313 \cr
}
}

\seealso{
Other Religion: 
\code{\link{churhact}},
\code{\link{churhmem}},
\code{\link{god}},
\code{\link{grace16}},
\code{\link{join16}},
\code{\link{lapsed}},
\code{\link{reborn}},
\code{\link{savesoul}},
\code{\link{saygrace}}
}
\concept{Religion}
\keyword{variable}
