% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_151.R
\name{tvshows}
\alias{tvshows}
\title{How often r watches tv drama or sitcoms}
\source{
General Social Survey https://gss.norc.org
}
\description{
tvshows
}
\details{
Question I'll now ask about some different kinds of television shows. Would you tell me how often you watch prime-time drama or situation comedy programs?  Would you say every day, several times a week, several times a month, rarely, or never?
A. Prime-time drama or situation comedy programs.
}
\section{Values}{

\itemize{
\item \code{1} daily
\item \code{2} sevrl days in week
\item \code{3} sevrl days in month
\item \code{4} rarely
\item \code{5} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1419/vshow}{GSS Data Explorer page for \code{tvshows}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab daily \tab sevrl days in week \tab sevrl days in month \tab rarely \tab never \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 336 \tab 587 \tab 275 \tab 296 \tab 101 \tab 3 \tab 8 \tab 1606 \cr
   Total \tab 0 \tab 336 \tab 587 \tab 275 \tab 296 \tab 101 \tab 3 \tab 8 \tab 1606 \cr
}
}

\seealso{
Other Culture: 
\code{\link{classics}},
\code{\link{english}},
\code{\link{grtbooks}},
\code{\link{judgeart}},
\code{\link{modpaint}},
\code{\link{pclit}},
\code{\link{trstprof}},
\code{\link{tvnews}},
\code{\link{tvpbs}}
}
\concept{Culture}
\keyword{variable}
