% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_151.R
\name{tvshows}
\alias{tvshows}
\title{How often r watches tv drama or sitcoms}
\source{
General Social Survey https://gss.norc.org
}
\description{
tvshows
}
\details{
Question I'll now ask about some different kinds of television shows. Would you tell me how often you watch prime-time drama or situation comedy programs?  Would you say every day, several times a week, several times a month, rarely, or never?
A. Prime-time drama or situation comedy programs.
}
\section{Values}{

\itemize{
\item \code{1} daily
\item \code{2} sevrl days in week
\item \code{3} sevrl days in month
\item \code{4} rarely
\item \code{5} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1419/vshow}{GSS Data Explorer page for \code{tvshows}}.
}

\seealso{
Other Culture: 
\code{\link{classics}},
\code{\link{english}},
\code{\link{grtbooks}},
\code{\link{judgeart}},
\code{\link{modpaint}},
\code{\link{pclit}},
\code{\link{trstprof}},
\code{\link{tvnews}},
\code{\link{tvpbs}}

Other Television Viewing: 
\code{\link{tvnews}},
\code{\link{tvpbs}}
}
\concept{Culture}
\concept{Television Viewing}
\keyword{variable}
