% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_125.R
\name{twoclass}
\alias{twoclass}
\title{Weak vs. strong}
\source{
General Social Survey https://gss.norc.org
}
\description{
twoclass
}
\details{
Question Now I am going to read you several statements. For each one please tell me whether you agree strongly, agree, disagree, or disagree strongly.
C. People can be divided into two classes--the weak and the strong
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1168/vshow}{GSS Data Explorer page for \code{twoclass}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 68 \tab 503 \tab 605 \tab 132 \tab 60 \tab 4 \tab 1372 \cr
   Total \tab 0 \tab 68 \tab 503 \tab 605 \tab 132 \tab 60 \tab 4 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{openmind}},
\code{\link{ownthing}},
\code{\link{talkback}}
}
\concept{Intergroup Relations}
\keyword{variable}
