% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_488.R
\name{twoincs1}
\alias{twoincs1}
\title{Both men and women should contribute to income}
\source{
General Social Survey https://gss.norc.org
}
\description{
twoincs1
}
\details{
Question A. Do you agree or disagree ... both the husband and the wife should contribute to the household income.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2012 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4455/vshow}{GSS Data Explorer page for \code{twoincs1}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{earnshh}},
\code{\link{fewknokd}},
\code{\link{happy7}},
\code{\link{kidfinbu}},
\code{\link{kidjob}},
\code{\link{satfam7}},
\code{\link{satjob7}},
\code{\link{ssfchild}},
\code{\link{ssmchild}}

Other Women: 
\code{\link{fewknokd}}
}
\concept{ISSP Family & Gender Roles}
\concept{Women}
\keyword{variable}
