% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_480.R
\name{twokids}
\alias{twokids}
\title{Desirability of two-kid family}
\source{
General Social Survey https://gss.norc.org
}
\description{
twokids
}
\details{
Question In general, what do you feel about each of these family sizes?
C. A family with two children
}
\section{Values}{

\itemize{
\item \code{1} very desirable
\item \code{2} desirable
\item \code{3} neither
\item \code{4} undesirable
\item \code{5} very undesirable
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4377/vshow}{GSS Data Explorer page for \code{twokids}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{fourkids}},
\code{\link{kidtrble}},
\code{\link{mardiv}},
\code{\link{nokids}},
\code{\link{numkids}},
\code{\link{onekid}},
\code{\link{threkids}}

Other Children: 
\code{\link{fourkids}},
\code{\link{kidtrble}},
\code{\link{nokids}},
\code{\link{numkids}},
\code{\link{onekid}},
\code{\link{threkids}}
}
\concept{Children}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
