% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_191.R
\name{typdealr}
\alias{typdealr}
\title{Rwas the agency a new or used car dealer'ship?}
\source{
General Social Survey https://gss.norc.org
}
\description{
typdealr
}
\details{
Question Was the agency a new car dealership or did it only sell used cars?
}
\section{Values}{

\itemize{
\item \code{1} new car dealership
\item \code{2} used car dealership
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1801/vshow}{GSS Data Explorer page for \code{typdealr}}.
}

\seealso{
Other Markets: 
\code{\link{cardealr}},
\code{\link{carprivt}},
\code{\link{newused}},
\code{\link{relprivt}},
\code{\link{warrntyc}}
}
\concept{Markets}
\keyword{variable}
