% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_238.R
\name{typeorg}
\alias{typeorg}
\title{Work in independent or part of larger org}
\source{
General Social Survey https://gss.norc.org
}
\description{
typeorg
}
\details{
Question I would like to ask you a few questions about the place you worked full or part time last week. Is the place part of a larger organization or is it completely independent in finance and personnel?
}
\section{Values}{

\itemize{
\item \code{1} larger organiztn
\item \code{2} cmplte indpndent
\item \code{8} not sure
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2140/vshow}{GSS Data Explorer page for \code{typeorg}}.
}

\seealso{
Other Jobs: 
\code{\link{occtrain}},
\code{\link{occyrs}},
\code{\link{trainsch}}

Other Job: 
\code{\link{occtrain}},
\code{\link{occyrs}},
\code{\link{trainsch}}
}
\concept{Job}
\concept{Jobs}
\keyword{variable}
