% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_426.R
\name{typevisa}
\alias{typevisa}
\title{What kind of visa}
\source{
General Social Survey https://gss.norc.org
}
\description{
typevisa
}
\details{
Question What kind of visa?
}
\section{Values}{

\itemize{
\item \code{1} h1b
\item \code{2} l1
\item \code{3} student
\item \code{4} tourist
\item \code{5} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3901/vshow}{GSS Data Explorer page for \code{typevisa}}.
}

\seealso{
Other Global Economics: 
\code{\link{agecmeus}},
\code{\link{bornsp}},
\code{\link{emphlth}},
\code{\link{emphplan}},
\code{\link{natdeg}},
\code{\link{othplan}},
\code{\link{visa}}

Other Citizenship: 
\code{\link{agecmeus}},
\code{\link{bornsp}},
\code{\link{visa}}

Other Immigration: 
\code{\link{agecmeus}},
\code{\link{bornsp}},
\code{\link{natdeg}},
\code{\link{visa}}
}
\concept{Citizenship}
\concept{Global Economics}
\concept{Immigration}
\keyword{variable}
