% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_426.R
\name{typevisa}
\alias{typevisa}
\title{What kind of visa}
\source{
General Social Survey https://gss.norc.org
}
\description{
typevisa
}
\details{
Question What kind of visa?
}
\section{Values}{

\itemize{
\item \code{1} h1b
\item \code{2} l1
\item \code{3} student
\item \code{4} tourist
\item \code{5} other
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3901/vshow}{GSS Data Explorer page for \code{typevisa}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab h1b \tab l1 \tab student \tab tourist \tab other \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1982 \tab 6 \tab 2 \tab 4 \tab 2 \tab 22 \tab 4 \tab 1 \tab 2023 \cr
   Total \tab 1982 \tab 6 \tab 2 \tab 4 \tab 2 \tab 22 \tab 4 \tab 1 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{agecmeus}},
\code{\link{bornsp}},
\code{\link{emphlth}},
\code{\link{emphplan}},
\code{\link{natdeg}},
\code{\link{othplan}},
\code{\link{visa}}
}
\concept{Global Economics}
\keyword{variable}
