% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_426.R
\name{typevisa}
\alias{typevisa}
\title{What kind of visa}
\source{
General Social Survey https://gss.norc.org
}
\description{
typevisa
}
\details{
Question What kind of visa?
}
\section{Values}{

\itemize{
\item \verb{[1]} h1b
\item \verb{[2]} l1
\item \verb{[3]} student
\item \verb{[4]} tourist
\item \verb{[5]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3901/vshow}{GSS Data Explorer page for \code{typevisa}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab h1b \tab l1 \tab student \tab tourist \tab other \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1982 \tab 6 \tab 2 \tab 4 \tab 2 \tab 22 \tab 4 \tab 1 \tab 2023 \cr
   Total \tab 1982 \tab 6 \tab 2 \tab 4 \tab 2 \tab 22 \tab 4 \tab 1 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{agecmeus}},
\code{\link{bornsp}},
\code{\link{emphlth}},
\code{\link{emphplan}},
\code{\link{natdeg}},
\code{\link{othplan}},
\code{\link{visa}}
}
\concept{Global Economics}
\keyword{variable}
