% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_348.R
\name{typinfrm}
\alias{typinfrm}
\title{R belongs to more than one informal groups}
\source{
General Social Survey https://gss.norc.org
}
\description{
typinfrm
}
\details{
Question Q. Do you belong to more than one informal group?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3126/vshow}{GSS Data Explorer page for \code{typinfrm}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2412 \tab 94 \tab 295 \tab 1 \tab 10 \tab 2812 \cr
   Total \tab 2412 \tab 94 \tab 295 \tab 1 \tab 10 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{numfrat}},
\code{\link{numpolit}},
\code{\link{numserv}},
\code{\link{numsport}},
\code{\link{numunion}},
\code{\link{numvet}},
\code{\link{numyouth}},
\code{\link{typchurh}},
\code{\link{typother}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
