% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_348.R
\name{typother}
\alias{typother}
\title{R belongs to more than one other groups}
\source{
General Social Survey https://gss.norc.org
}
\description{
typother
}
\details{
Question P. Do you belong to more than one of any other club?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3125/vshow}{GSS Data Explorer page for \code{typother}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2004 \tab 2706 \tab 9 \tab 88 \tab 9 \tab 2812 \cr
   Total \tab 2706 \tab 9 \tab 88 \tab 9 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{numfrat}},
\code{\link{numpolit}},
\code{\link{numserv}},
\code{\link{numsport}},
\code{\link{numunion}},
\code{\link{numvet}},
\code{\link{numyouth}},
\code{\link{typchurh}},
\code{\link{typinfrm}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
