% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_82.R
\name{unemp5}
\alias{unemp5}
\title{Unemployed in last 5 yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
unemp5
}
\details{
Question Unemployment during last five years.
}
\section{Values}{

\itemize{
\item \code{0} no
\item \code{1} 4 prior yrs
\item \code{2} 4 prior yrs, earner
\item \code{3} last yr
\item \code{4} both,not earner
\item \code{5} both,earner prior 4
\item \code{6} last yr, earner
\item \code{7} both,earner last yr
\item \code{8} both,earner both
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1978 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab B/C \tab partial \cr
   1989 \tab B/C \tab partial \cr
   1990 \tab B/C \tab partial \cr
   1991 \tab B/C \tab partial \cr
   1993 \tab B/C \tab partial \cr
   1994 \tab B/C \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/740/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab 4 prior yrs \tab 4 prior yrs, earner \tab both,earner both \tab both,earner last yr \tab both,earner prior 4 \tab both,not earner \tab last yr \tab last yr, earner \tab no \tab no answer \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 76 \tab 92 \tab 58 \tab 10 \tab 4 \tab 47 \tab 23 \tab 22 \tab 1193 \tab 7 \tab - \tab 1532 \cr
   1980 \tab - \tab 79 \tab 88 \tab 42 \tab 7 \tab 1 \tab 41 \tab 16 \tab 33 \tab 1157 \tab 4 \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab - \tab 82 \tab 75 \tab 99 \tab 6 \tab 9 \tab 59 \tab 35 \tab 40 \tab 1184 \tab 10 \tab - \tab 1599 \cr
   1984 \tab - \tab 65 \tab 95 \tab 79 \tab 10 \tab 5 \tab 61 \tab 19 \tab 26 \tab 1097 \tab 16 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab - \tab 87 \tab 71 \tab 65 \tab 10 \tab 7 \tab 44 \tab 20 \tab 19 \tab 1138 \tab 9 \tab - \tab 1470 \cr
   1987 \tab - \tab 106 \tab 118 \tab 78 \tab 7 \tab 6 \tab 56 \tab 28 \tab 34 \tab 1371 \tab 15 \tab - \tab 1819 \cr
   1988 \tab 484 \tab 72 \tab 67 \tab 37 \tab 4 \tab 3 \tab 24 \tab 16 \tab 16 \tab 753 \tab 5 \tab - \tab 1481 \cr
   1989 \tab 502 \tab 47 \tab 55 \tab 28 \tab 2 \tab 4 \tab 17 \tab 16 \tab 16 \tab 840 \tab 10 \tab - \tab 1537 \cr
   1990 \tab 473 \tab 37 \tab 50 \tab 39 \tab 3 \tab 5 \tab 20 \tab 9 \tab 14 \tab 718 \tab 4 \tab - \tab 1372 \cr
   1991 \tab 500 \tab 53 \tab 46 \tab 46 \tab 3 \tab 3 \tab 30 \tab 12 \tab 15 \tab 799 \tab 10 \tab - \tab 1517 \cr
   1993 \tab 549 \tab 56 \tab 61 \tab 40 \tab 2 \tab 7 \tab 38 \tab 16 \tab 20 \tab 814 \tab 3 \tab - \tab 1606 \cr
   1994 \tab 2481 \tab 17 \tab 29 \tab 29 \tab 4 \tab 3 \tab 16 \tab 4 \tab 6 \tab 398 \tab 5 \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 4510 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 53969 \tab 777 \tab 847 \tab 640 \tab 68 \tab 57 \tab 453 \tab 214 \tab 261 \tab 11462 \tab 98 \tab 6853 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{divorce5}},
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{racdif4}},
\code{\link{salabort}},
\code{\link{salfergt}},
\code{\link{sallabor}},
\code{\link{salsatfn}}

Other Trauma: 
\code{\link{divorce5}},
\code{\link{divrel1}},
\code{\link{divrel4}}
}
\concept{Core}
\concept{Employment}
\concept{Trauma}
\concept{Unemployment}
\keyword{variable}
