% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_82.R
\name{unemp5}
\alias{unemp5}
\title{Unemployed in last 5 yrs}
\source{
General Social Survey https://gss.norc.org
}
\description{
unemp5
}
\details{
Question Unemployment during last five years.
}
\section{Values}{

\itemize{
\item \code{0} no
\item \code{1} 4 prior yrs
\item \code{2} 4 prior yrs, earner
\item \code{3} last yr
\item \code{4} both,not earner
\item \code{5} both,earner prior 4
\item \code{6} last yr, earner
\item \code{7} both,earner last yr
\item \code{8} both,earner both
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1978 \tab -/- \tab full \cr
   1980 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1987 \tab -/- \tab full \cr
   1988 \tab B/C \tab partial \cr
   1989 \tab B/C \tab partial \cr
   1990 \tab B/C \tab partial \cr
   1991 \tab B/C \tab partial \cr
   1993 \tab B/C \tab partial \cr
   1994 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/740/vshow}{GSS Data Explorer page for \code{unemp5}}.
}

\seealso{
Other Core: 
\code{\link{divorce5}},
\code{\link{divrel1}},
\code{\link{divrel4}},
\code{\link{racdif4}},
\code{\link{salabort}},
\code{\link{salfergt}},
\code{\link{sallabor}},
\code{\link{salsatfn}},
\code{\link{salsci}}

Other Trauma: 
\code{\link{divorce5}},
\code{\link{divrel1}},
\code{\link{divrel4}}
}
\concept{Core}
\concept{Employment}
\concept{Trauma}
\concept{Unemployment}
\keyword{variable}
