% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_501.R
\name{unhlthy}
\alias{unhlthy}
\title{How often does r work in unhealthy conds}
\source{
General Social Survey https://gss.norc.org
}
\description{
unhlthy
}
\details{
Question Now some more questions about your working conditions. one code for each item below to show how often it applies to your work. How often...
F. Do you work in unhealthy conditions?
}
\section{Values}{

\itemize{
\item \code{1} always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} hardly ever
\item \code{5} never
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4555/vshow}{GSS Data Explorer page for \code{unhlthy}}.
}

\seealso{
Other Work Orientation: 
\code{\link{boredom}},
\code{\link{bossemps}},
\code{\link{cowrkers}},
\code{\link{danger}},
\code{\link{findwork}},
\code{\link{jobsat}},
\code{\link{planwrk}},
\code{\link{stress}},
\code{\link{unpleznt}}

Other Job: 
\code{\link{boredom}},
\code{\link{bossemps}},
\code{\link{cowrkers}},
\code{\link{danger}},
\code{\link{jobsat}},
\code{\link{planwrk}},
\code{\link{stress}},
\code{\link{unpleznt}}
}
\concept{Job}
\concept{Work Orientation}
\keyword{variable}
