% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_65.R
\name{union}
\alias{union}
\title{Does r or spouse belong to union}
\source{
General Social Survey https://gss.norc.org
}
\description{
union
}
\details{
Question
}
\section{Values}{

\itemize{
\item \code{1} yes, respondent belongs
\item \code{2} yes, spouse belongs
\item \code{3} yes, both belong
\item \code{4} no, neither belong
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1976 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1980 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1986 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab B/C/- \tab partial \cr
   1989 \tab B/C/- \tab partial \cr
   1990 \tab B/C/- \tab partial \cr
   1991 \tab A/B/C \tab full \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab B/C/- \tab partial \cr
   1996 \tab B/C/- \tab partial \cr
   1998 \tab B/C/- \tab partial \cr
   2000 \tab B/C/- \tab partial \cr
   2002 \tab B/C/- \tab partial \cr
   2004 \tab B/C/- \tab partial \cr
   2006 \tab B/C/D \tab partial \cr
   2008 \tab B/C/- \tab partial \cr
   2010 \tab B/C/- \tab partial \cr
   2012 \tab B/C/- \tab partial \cr
   2014 \tab B/C/- \tab partial \cr
   2016 \tab B/C/- \tab partial \cr
   2018 \tab B/C/- \tab partial \cr
   2021 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/584/vshow}{GSS Data Explorer page for \code{union}}.
}

\seealso{
Other Core: 
\code{\link{getahead}},
\code{\link{getaid}},
\code{\link{govaid}},
\code{\link{parsol}},
\code{\link{unemp}},
\code{\link{union1}},
\code{\link{wksup1}},
\code{\link{wksups}},
\code{\link{wksups1}}
}
\concept{Core}
\keyword{variable}
