% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_445.R
\name{unionsok}
\alias{unionsok}
\title{How good union for the country}
\source{
General Social Survey https://gss.norc.org
}
\description{
unionsok
}
\details{
Question In general, how good would you say labor unions are for the country as a whole?
}
\section{Values}{

\itemize{
\item \code{1} excellent
\item \code{2} very good
\item \code{3} fairly good
\item \code{4} not very good
\item \code{5} not good at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4088/vshow}{GSS Data Explorer page for \code{unionsok}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab excellent \tab very good \tab fairly good \tab not very good \tab not good at all \tab can't choose \tab no answer \tab Total \cr
   1990 \tab 155 \tab 33 \tab 183 \tab 593 \tab 209 \tab 63 \tab 108 \tab 28 \tab 1372 \cr
   Total \tab 155 \tab 33 \tab 183 \tab 593 \tab 209 \tab 63 \tab 108 \tab 28 \tab 1372 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{gendereq}},
\code{\link{grnlaws}},
\code{\link{polint}},
\code{\link{prvdhlth}},
\code{\link{prvdold}},
\code{\link{rprtst1}},
\code{\link{rprtst15}},
\code{\link{rprtst3}},
\code{\link{rprtst35}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
