% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_147.R
\name{united}
\alias{united}
\title{Americans united or divided on values}
\source{
General Social Survey https://gss.norc.org
}
\description{
united
}
\details{
Question A. There is a lot of discussion today about whether Americans are divided or united. Some say that Americans are united and in agreement about the most important values. Others think that Americans are greatly divided when it comes to the most important values. What is your view about this?
}
\section{Values}{

\itemize{
\item \verb{[1]} american are united
\item \verb{[2]} american are divided
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1380/vshow}{GSS Data Explorer page for \code{united}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab american are united \tab american are divided \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 518 \tab 1009 \tab 73 \tab 6 \tab 1606 \cr
   1994 \tab 1518 \tab 581 \tab 810 \tab 73 \tab 10 \tab 2992 \cr
   Total \tab 1518 \tab 1099 \tab 1819 \tab 146 \tab 16 \tab 4598 \cr
}
}

\seealso{
Other Culture: 
\code{\link{bigband}},
\code{\link{blugrass}},
\code{\link{lfechnce}},
\code{\link{lfegenes}},
\code{\link{lfegod}},
\code{\link{lfehrdwk}},
\code{\link{lfesocty}},
\code{\link{obtohelp}},
\code{\link{opoutcme}}
}
\concept{Culture}
\concept{Multiculturalism}
\keyword{variable}
