% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_645.R
\name{univgndr}
\alias{univgndr}
\title{Man vs woman head of university}
\source{
General Social Survey https://gss.norc.org
}
\description{
univgndr
}
\details{
Question (Please consider the following positions and select the options that come closest to your own opinion.) Head of a university
}
\section{Values}{

\itemize{
\item \code{1} women are much better suited
\item \code{2} women are somewhat better suited
\item \code{3} women and men are equally suited
\item \code{4} men are somewhat better suited
\item \code{5} men are much better suited
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7784/vshow}{GSS Data Explorer page for \code{univgndr}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab women are much better suited \tab women are somewhat better suited \tab women and men are equally suited \tab men are somewhat better suited \tab men are much better suited \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2386 \tab 36 \tab 32 \tab 936 \tab 39 \tab 18 \tab 84 \tab 5 \tab 8 \tab 3544 \cr
   Total \tab 2386 \tab 36 \tab 32 \tab 936 \tab 39 \tab 18 \tab 84 \tab 5 \tab 8 \tab 3544 \cr
}
}

\keyword{variable}
