% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_145.R
\name{unmanrel}
\alias{unmanrel}
\title{Relations between management and workers}
\source{
General Social Survey https://gss.norc.org
}
\description{
unmanrel
}
\details{
Question In general how would you describe relations in your work place between management and employees?  Would you say they are very good, quite good, neither good nor bad, quite bad, or very bad?
}
\section{Values}{

\itemize{
\item \code{1} very
\item \code{2} quite good
\item \code{3} neither good nor bad
\item \code{4} quite bad
\item \code{5} very bad
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1366/vshow}{GSS Data Explorer page for \code{unmanrel}}.
}

\seealso{
Other Work Organizations: 
\code{\link{cowrkrel}},
\code{\link{fringe10}},
\code{\link{fringe11}},
\code{\link{fringe7}},
\code{\link{fringe8}},
\code{\link{fringe9}},
\code{\link{jobvshme}},
\code{\link{schoolng}},
\code{\link{unvote}}
}
\concept{Work Organizations}
\keyword{variable}
