% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_501.R
\name{unpleznt}
\alias{unpleznt}
\title{How often does r work in unpleasant conds}
\source{
General Social Survey https://gss.norc.org
}
\description{
unpleznt
}
\details{
Question Now some more questions about your working conditions. Please circle one code for each item below to show how often it applies to your work. How often...
G. Do you work in physically unpleasant conditions?
}
\section{Values}{

\itemize{
\item \code{1} always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} hardly ever
\item \code{5} never
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4556/vshow}{GSS Data Explorer page for \code{unpleznt}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab always \tab often \tab sometimes \tab hardly ever \tab never \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 639 \tab 14 \tab 50 \tab 179 \tab 220 \tab 405 \tab 6 \tab 24 \tab 1537 \cr
   Total \tab 639 \tab 14 \tab 50 \tab 179 \tab 220 \tab 405 \tab 6 \tab 24 \tab 1537 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{boredom}},
\code{\link{bossemps}},
\code{\link{cowrkers}},
\code{\link{danger}},
\code{\link{findwork}},
\code{\link{jobsat}},
\code{\link{planwrk}},
\code{\link{stress}},
\code{\link{unhlthy}}
}
\concept{Work Orientation}
\keyword{variable}
