% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_674.R
\name{unplgrsn}
\alias{unplgrsn}
\title{Reason r reduced time spent on internet}
\source{
General Social Survey https://gss.norc.org
}
\description{
unplgrsn
}
\details{
Question What is the main reason why you deliberately reduced the time you spent on the Internet?
}
\section{Values}{

\itemize{
\item \code{1} i had concerns about data security or privacy issues
\item \code{2} i felt it negatively impacted my physical or mental wellbeing
\item \code{3} i felt overwhelmed by the amount of information
\item \code{4} i was asked to spend more time in person with family or friends
\item \code{5} i wanted more time for offline activities
\item \code{6} i couldn't afford to use the internet
\item \code{7} other reasons
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8179/vshow}{GSS Data Explorer page for \code{unplgrsn}}.
}

\seealso{
Other ISSP Digital Societies: 
\code{\link{intapps}},
\code{\link{intbhlf}},
\code{\link{intcomm}},
\code{\link{intliteracy}},
\code{\link{intpub}},
\code{\link{intshare}},
\code{\link{intskill}},
\code{\link{intsurf}},
\code{\link{unplug}}
}
\concept{ISSP Digital Societies}
\keyword{variable}
