% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_674.R
\name{unplgrsn}
\alias{unplgrsn}
\title{Reason r reduced time spent on internet}
\source{
General Social Survey https://gss.norc.org
}
\description{
unplgrsn
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} i had concerns about data security or privacy issues
\item \code{2} i felt it negatively impacted my physical or mental wellbeing
\item \code{3} i felt overwhelmed by the amount of information
\item \code{4} i was asked to spend more time in person with family or friends
\item \code{5} i wanted more time for offline activities
\item \code{6} i couldn't afford to use the internet
\item \code{7} other reasons
\item \code{NA(d)} can't choose
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
