% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_135.R
\name{unpower}
\alias{unpower}
\title{Unions in this country have too little power}
\source{
General Social Survey https://gss.norc.org
}
\description{
unpower
}
\details{
Question To what extent do you agree or disagree with the following statements?
Please tell me whether you strongly agree, agree, disagree, or strongly disagree:
C. Unions in this country have too little power…
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1265/vshow}{GSS Data Explorer page for \code{unpower}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   1991 \tab - \tab 87 \tab 336 \tab 721 \tab 203 \tab 165 \tab 5 \tab 1517 \cr
   Total \tab 0 \tab 87 \tab 336 \tab 721 \tab 203 \tab 165 \tab 5 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{genecanx}},
\code{\link{genecany}},
\code{\link{genegets}},
\code{\link{merged}},
\code{\link{outofbiz}},
\code{\link{profits1}},
\code{\link{profits2}},
\code{\link{reorg}},
\code{\link{unprog}}
}
\concept{Work Organizations}
\keyword{variable}
