% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_83.R
\name{unrel1}
\alias{unrel1}
\title{Relatives unemployed in last yr}
\source{
General Social Survey https://gss.norc.org
}
\description{
unrel1
}
\details{
Question A. Number of relatives unemployed during last year.
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1978, 1980, 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/741/vshow}{GSS Data Explorer page for \code{unrel1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab Total \cr
   1978 \tab - \tab 1107 \tab 352 \tab 64 \tab 7 \tab 2 \tab - \tab 1532 \cr
   1980 \tab - \tab 1070 \tab 333 \tab 55 \tab 9 \tab 1 \tab - \tab 1468 \cr
   1983 \tab - \tab 955 \tab 480 \tab 127 \tab 26 \tab 9 \tab 2 \tab 1599 \cr
   1984 \tab - \tab 954 \tab 401 \tab 95 \tab 20 \tab 3 \tab - \tab 1473 \cr
   Total \tab 0 \tab 4086 \tab 1566 \tab 341 \tab 62 \tab 15 \tab 2 \tab 6072 \cr
}
}

\seealso{
Other Core: 
\code{\link{chlddth}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{hosdis5}},
\code{\link{hosrel1}},
\code{\link{hosrel4}},
\code{\link{madeath}},
\code{\link{padeath}},
\code{\link{unrel4}}
}
\concept{Core}
\keyword{variable}
