% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_83.R
\name{unrel4}
\alias{unrel4}
\title{Relatives unemp between 5 & 1 yrs ago}
\source{
General Social Survey https://gss.norc.org
}
\description{
unrel4
}
\details{
Question B. Number or relatives unemployed during previous four years.
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1978, 1980, 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/742/vshow}{GSS Data Explorer page for \code{unrel4}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 2 \tab 3 \tab 4 \tab 5 \tab Total \cr
   1978 \tab - \tab 1031 \tab 407 \tab 73 \tab 18 \tab 3 \tab - \tab 1532 \cr
   1980 \tab - \tab 1029 \tab 371 \tab 53 \tab 12 \tab 3 \tab - \tab 1468 \cr
   1983 \tab - \tab 1008 \tab 461 \tab 99 \tab 23 \tab 6 \tab 2 \tab 1599 \cr
   1984 \tab - \tab 923 \tab 444 \tab 85 \tab 18 \tab 3 \tab - \tab 1473 \cr
   Total \tab 0 \tab 3991 \tab 1683 \tab 310 \tab 71 \tab 15 \tab 2 \tab 6072 \cr
}
}

\seealso{
Other Core: 
\code{\link{chlddth}},
\code{\link{death16}},
\code{\link{death5}},
\code{\link{hosdis5}},
\code{\link{hosrel1}},
\code{\link{hosrel4}},
\code{\link{madeath}},
\code{\link{padeath}},
\code{\link{unrel1}}
}
\concept{Core}
\keyword{variable}
