% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_476.R
\name{unsklrnk}
\alias{unsklrnk}
\title{Rank in society of unskilled workers}
\source{
General Social Survey https://gss.norc.org
}
\description{
unsklrnk
}
\details{
Question Now we would like to know where you and other people fit in. Please look at the diagram you think best describes America today, the one you chose first:
B. Where would you say an unskilled worker in a factory fits in?
}
\section{Values}{

\itemize{
\item \code{1} top
\item \code{7} bottom
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4339/vshow}{GSS Data Explorer page for \code{unsklrnk}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab top \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab bottom \tab can't choose \tab no answer \tab Total \cr
   2000 \tab 1545 \tab 12 \tab 6 \tab 25 \tab 101 \tab 237 \tab 457 \tab 293 \tab 61 \tab 80 \tab 2817 \cr
   Total \tab 1545 \tab 12 \tab 6 \tab 25 \tab 101 \tab 237 \tab 457 \tab 293 \tab 61 \tab 80 \tab 2817 \cr
}
}

\seealso{
Other ISSP Social Inequality: 
\code{\link{execrnk}},
\code{\link{famrnk}},
\code{\link{justpay}},
\code{\link{pasup}},
\code{\link{paychild}},
\code{\link{paydowel}},
\code{\link{payhard}},
\code{\link{soctype1}},
\code{\link{soctype2}}
}
\concept{ISSP Social Inequality}
\keyword{variable}
