% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_497.R
\name{unslfcon}
\alias{unslfcon}
\title{Rank of loss of confide. when unemployed}
\source{
General Social Survey https://gss.norc.org
}
\description{
unslfcon
}
\details{
Question Suppose you were unemployed and couldn't find a job. Which of the following problems would be the worst?  Please write '1' in the box beside the worst thing, write '2' in the box beside the next worst thing, and '3' beside the third worst thing. Leave the other boxes blank.
C. Loss of self-confidence
}
\section{Values}{

\itemize{
\item \code{1} worst
\item \code{2} 2nd worst
\item \code{3} 3rd worst
\item \code{4} not chosen
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1989 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4518/vshow}{GSS Data Explorer page for \code{unslfcon}}.
}

\seealso{
Other Work Orientation: 
\code{\link{absent}},
\code{\link{adforjob}},
\code{\link{ageemp}},
\code{\link{appemps}},
\code{\link{askhelp}},
\code{\link{boredom}},
\code{\link{bossemps}},
\code{\link{bosswrks}},
\code{\link{chngewrk}},
\code{\link{chngwork}},
\code{\link{cowrkers}},
\code{\link{dailywrk}},
\code{\link{danger}},
\code{\link{decidwrk}},
\code{\link{discwk5}},
\code{\link{dk}},
\code{\link{dowell}},
\code{\link{econsup}},
\code{\link{econsup1}},
\code{\link{edcskill}},
\code{\link{educemp}},
\code{\link{employer}},
\code{\link{empself}},
\code{\link{evjob}},
\code{\link{exjobsat}},
\code{\link{expernc}},
\code{\link{extraern}},
\code{\link{extrawrk}},
\code{\link{famintjb}},
\code{\link{famorjob}},
\code{\link{famresp}},
\code{\link{findjob}},
\code{\link{findwork}},
\code{\link{flexhrs}},
\code{\link{flexhrs1}},
\code{\link{getjob}},
\code{\link{harass5}},
\code{\link{helporg1}},
\code{\link{hiinc}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{hrsmoney}},
\code{\link{hwduties}},
\code{\link{indusoth}},
\code{\link{intjob}},
\code{\link{iwrkhard}},
\code{\link{jbendmo}},
\code{\link{jbendyr}},
\code{\link{jbintfam}},
\code{\link{jbtrain}},
\code{\link{jobeasy}},
\code{\link{joborfam}},
\code{\link{jobsat}},
\code{\link{jobskill}},
\code{\link{leavejob}},
\code{\link{leisure}},
\code{\link{localnum}},
\code{\link{lookjob}},
\code{\link{lowpay}},
\code{\link{moonlite}},
\code{\link{mortravl}},
\code{\link{moveaway}},
\code{\link{moveinus}},
\code{\link{newjob}},
\code{\link{newskill}},
\code{\link{nounemp1}},
\code{\link{nounemp2}},
\code{\link{nounemp3}},
\code{\link{nounemp4}},
\code{\link{nounemp5}},
\code{\link{nounemp6}},
\code{\link{numemply}},
\code{\link{numsites}},
\code{\link{otherwrk}},
\code{\link{othhrs}},
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paysame}},
\code{\link{paytime}},
\code{\link{physwrk}},
\code{\link{placewrk}},
\code{\link{planwrk}},
\code{\link{priagncy}},
\code{\link{prideorg}},
\code{\link{privgovt}},
\code{\link{promotn}},
\code{\link{proudwrk}},
\code{\link{pubagncy}},
\code{\link{rageemp}},
\code{\link{rcontact}},
\code{\link{rdk}},
\code{\link{rdowell}},
\code{\link{reducemp}},
\code{\link{replaceu}},
\code{\link{rexpernc}},
\code{\link{rfamresp}},
\code{\link{rflexhrs}},
\code{\link{rhiinc}},
\code{\link{rhlpoths}},
\code{\link{rhlpsoc}},
\code{\link{rimpskls}},
\code{\link{rintjob}},
\code{\link{rleisure}},
\code{\link{rpaysame}},
\code{\link{rpromotn}},
\code{\link{rsecjob}},
\code{\link{rwrkindp}},
\code{\link{ryrsfirm}},
\code{\link{secjob}},
\code{\link{selfemp1}},
\code{\link{selfemp2}},
\code{\link{sethours}},
\code{\link{sexemp}},
\code{\link{skltrain}},
\code{\link{smallbig}},
\code{\link{spwrkgvt}},
\code{\link{stayorg3}},
\code{\link{stress}},
\code{\link{strngun}},
\code{\link{supervis}},
\code{\link{supnum}},
\code{\link{suprvsjb}},
\code{\link{techjobs}},
\code{\link{techwork}},
\code{\link{tempwork}},
\code{\link{timefam}},
\code{\link{timefrnd}},
\code{\link{timehhwk}},
\code{\link{timeleis}},
\code{\link{timeoff}},
\code{\link{timepdwk}},
\code{\link{timerelx}},
\code{\link{totalnum}},
\code{\link{unbetter}},
\code{\link{unbored}},
\code{\link{undk}},
\code{\link{unfamten}},
\code{\link{unhlthy}},
\code{\link{unionsbd}},
\code{\link{unjobexp}},
\code{\link{unjobsec}},
\code{\link{unmoney}},
\code{\link{unpeople}},
\code{\link{unpleznt}},
\code{\link{unrespct}},
\code{\link{useskill}},
\code{\link{wantads}},
\code{\link{wantjob}},
\code{\link{wantjob1}},
\code{\link{wantjob2}},
\code{\link{whydisc5}},
\code{\link{whyjbend}},
\code{\link{wkathome}},
\code{\link{wkpersnl}},
\code{\link{wktenure}},
\code{\link{wkwkends}},
\code{\link{worecsup}},
\code{\link{worknow}},
\code{\link{workweek}},
\code{\link{wornojob}},
\code{\link{worryjob}},
\code{\link{wrkearn}},
\code{\link{wrkenjoy}},
\code{\link{wrkimp}},
\code{\link{wrkindp}},
\code{\link{wrkshift}},
\code{\link{xhaustn}},
\code{\link{yrjobend}},
\code{\link{yrsfirm}}
}
\concept{Work Orientation}
\keyword{variable}
