% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_145.R
\name{unvote}
\alias{unvote}
\title{Would you vote for or against having a union}
\source{
General Social Survey https://gss.norc.org
}
\description{
unvote
}
\details{
Question If an election were held with secret ballots, would you vote for or against having a union represent you?
}
\section{Values}{

\itemize{
\item \code{1} for
\item \code{2} against
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1365/vshow}{GSS Data Explorer page for \code{unvote}}.
}

\seealso{
Other Work Organizations: 
\code{\link{cowrkrel}},
\code{\link{fringe10}},
\code{\link{fringe11}},
\code{\link{fringe7}},
\code{\link{fringe8}},
\code{\link{fringe9}},
\code{\link{jobvshme}},
\code{\link{schoolng}},
\code{\link{unmanrel}}
}
\concept{Work Organizations}
\keyword{variable}
