% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_211.R
\name{upsdowns}
\alias{upsdowns}
\title{How likely x is experiencing normal ups-and-downs of life}
\source{
General Social Survey https://gss.norc.org
}
\description{
upsdowns
}
\details{
Question In your opinion, how likely is it that NAME is experiencing (READ A-E) -- very likely, somewhat likely, not very likely, or not at all likely? (IN PART E, READ ONLY THE PHRASE FOLLOWING THE LETTER OF THE VIGNETTE)
A. Part of the normal ups and downs of life
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1994/vshow}{GSS Data Explorer page for \code{upsdowns}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 425 \tab 462 \tab 305 \tab 211 \tab 17 \tab 24 \tab 2904 \cr
   2006 \tab 3073 \tab 444 \tab 479 \tab 285 \tab 180 \tab 34 \tab 15 \tab 4510 \cr
   2018 \tab 1175 \tab 294 \tab 420 \tab 253 \tab 163 \tab 14 \tab 29 \tab 2348 \cr
   Total \tab 5708 \tab 1163 \tab 1361 \tab 843 \tab 554 \tab 65 \tab 68 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{breakdwn}},
\code{\link{decmoney}},
\code{\link{dectreat}},
\code{\link{genetics}},
\code{\link{godswill}},
\code{\link{imprvown}},
\code{\link{mentlill}},
\code{\link{physill}},
\code{\link{viglabel}}
}
\concept{Mental Health}
\keyword{variable}
