% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_211.R
\name{upsdowns}
\alias{upsdowns}
\title{How likely x is experiencing normal ups-and-downs of life}
\source{
General Social Survey https://gss.norc.org
}
\description{
upsdowns
}
\details{
Question In your opinion, how likely is it that NAME is experiencing (READ A-E) -- very likely, somewhat likely, not very likely, or not at all likely? (IN PART E, READ ONLY THE PHRASE FOLLOWING THE LETTER OF THE VIGNETTE)
A. Part of the normal ups and downs of life
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2018 \tab B/C/- \tab partial \cr
   2024 \tab A/B/- \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1994/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab don't know \tab no answer \tab not at all likely \tab not very likely \tab somewhat likely \tab very likely \tab not available in this year \tab skipped on web \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1460 \tab 17 \tab 24 \tab 211 \tab 305 \tab 462 \tab 425 \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 3073 \tab 34 \tab 15 \tab 180 \tab 285 \tab 479 \tab 444 \tab - \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 1175 \tab 14 \tab 29 \tab 163 \tab 253 \tab 420 \tab 294 \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab - \tab 3544 \cr
   2024 \tab 1691 \tab 21 \tab 10 \tab 196 \tab 372 \tab 643 \tab 357 \tab - \tab 19 \tab 3309 \cr
   Total \tab 66483 \tab 86 \tab 78 \tab 750 \tab 1215 \tab 2004 \tab 1520 \tab 3544 \tab 19 \tab 75699 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{breakdwn}},
\code{\link{decmoney}},
\code{\link{dectreat}},
\code{\link{genetics}},
\code{\link{godswill}},
\code{\link{imprvown}},
\code{\link{mentlill}},
\code{\link{physill}},
\code{\link{viglabel}}

Other Mental Health: 
\code{\link{breakdwn}},
\code{\link{decmoney}},
\code{\link{dectreat}},
\code{\link{genetics}},
\code{\link{godswill}},
\code{\link{imprvown}},
\code{\link{mentlill}},
\code{\link{physill}},
\code{\link{viglabel}}
}
\concept{Mental Health}
\keyword{variable}
