% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_256.R
\name{usamind}
\alias{usamind}
\title{What percentage of population is american india}
\source{
General Social Survey https://gss.norc.org
}
\description{
usamind
}
\details{
Question Just your best guess–what percentage of the United States population is each group? (INSTRUCTION: PERCENTAGES DO NOT HAVE TO ADD TO 100pct AND THE LISTED GROUPS MAY OVERLAP. ENCOURAGE ESTIMATES FOR ALL GROUPS, BUT DISCOURAGE RESPONDENTS REVISING THEIR ESTIMATES.)
F. American Indians?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2315/vshow}{GSS Data Explorer page for \code{usamind}}.
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{blkchng}},
\code{\link{hispchng}},
\code{\link{jewschng}},
\code{\link{usasn}},
\code{\link{usblk}},
\code{\link{ushisp}},
\code{\link{usjews}},
\code{\link{usmixed}},
\code{\link{whtchng}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
