% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_576.R
\name{uscitzn}
\alias{uscitzn}
\title{Is r us citizen}
\source{
General Social Survey https://gss.norc.org
}
\description{
uscitzn
}
\details{
Question If not born in US, is R a US citizen
As of 2024, USCITZN has been recoded to simplify responses. It is now a synthetic variable which represents the original USCITZN, as well as the experimental USCITZNV and USCITZNNV. This recode is retroactive, covering all previous years of USCITZN. The original data have been preserved in USCITZN_0822. USCITZNV and USCITZNNV will not be made available in 2024.
}
\section{Values}{

\itemize{
\item \code{1} a u.s. citizen
\item \code{2} not a u.s. citizen
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5283/vshow}{GSS Data Explorer page for \code{uscitzn}}.
}

\seealso{
Other Ethnicity: 
\code{\link{difrace1}},
\code{\link{difrace2}},
\code{\link{difrace3}},
\code{\link{eth1}},
\code{\link{eth2}},
\code{\link{eth3}},
\code{\link{ethnic}},
\code{\link{ethnum}},
\code{\link{fucitzn}},
\code{\link{hispanic}},
\code{\link{racecen1}},
\code{\link{racecen1_0022}},
\code{\link{racecen2}},
\code{\link{racecen2_0022}},
\code{\link{racecen3}},
\code{\link{racecen3_0022}},
\code{\link{racedbtf}},
\code{\link{racesee}},
\code{\link{raceself}},
\code{\link{speth1}},
\code{\link{speth2}},
\code{\link{speth3}},
\code{\link{spethnic}},
\code{\link{spethnum}},
\code{\link{uscitzn_0822}}
}
\concept{Ethnicity}
\keyword{variable}
