% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_686.R
\name{uscitznv_2122}
\alias{uscitznv_2122}
\title{If not born in us, is r a us citizen (volunteered option)}
\source{
General Social Survey https://gss.norc.org
}
\description{
uscitznv_2122
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} a u.s. citizen
\item \code{2} not a u.s. citizen
\item \code{3} a u.s. citizen born in puerto rico, the u.s. virgin islands, or the northern marianas islands (phone mode only: v
\item \code{4} born outside of the u.s. to parents who were u.s. citizens at that time (phone mode only: volunteered)
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
