% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_29.R
\name{usclass4}
\alias{usclass4}
\title{Ones own efforts don't count}
\source{
General Social Survey https://gss.norc.org
}
\description{
usclass4
}
\details{
Question Here are different opinions about social differences in this country. Please tell me for each one whether you strongly agree, somewhat, agree, somewhat disagree, or strongly disagree. READ EACH STATEMENT. CIRCLE ONE CODE FOR EACH.
D. What one gets in life hardly depends at all on one's own efforts, but rather on the economic situation, job opportunities, union agreements, and the social services provided by the government.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} somewhat agree
\item \code{3} somewhat disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/236/vshow}{GSS Data Explorer page for \code{usclass4}}.
}

\seealso{
Other Core: 
\code{\link{equal7}},
\code{\link{equal8}},
\code{\link{usclass1}},
\code{\link{usclass2}},
\code{\link{usclass3}},
\code{\link{usclass5}},
\code{\link{usclass6}},
\code{\link{usclass7}},
\code{\link{usclass8}}

Other Social Class: 
\code{\link{equal7}},
\code{\link{equal8}},
\code{\link{usclass1}},
\code{\link{usclass2}},
\code{\link{usclass3}},
\code{\link{usclass5}},
\code{\link{usclass6}},
\code{\link{usclass7}},
\code{\link{usclass8}}

Other Social Mobility: 
\code{\link{equal8}},
\code{\link{usclass2}},
\code{\link{usclass3}},
\code{\link{usclass7}}
}
\concept{Core}
\concept{Social Class}
\concept{Social Mobility}
\keyword{variable}
