% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_29.R
\name{usclass5}
\alias{usclass5}
\title{Personal income not determined by work}
\source{
General Social Survey https://gss.norc.org
}
\description{
usclass5
}
\details{
Question Here are different opinions about social differences in this country. Please tell me for each one whether you strongly agree, somewhat, agree, somewhat disagree, or strongly disagree. READ EACH STATEMENT. CIRCLE ONE CODE FOR EACH.
E. Personal income should not be determined solely by one's work. Rather, everybody should get what he/she needs to provide a decent life for his/her family.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} somewhat agree
\item \code{3} somewhat disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1984 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/237/vshow}{GSS Data Explorer page for \code{usclass5}}.
}

\seealso{
Other Core: 
\code{\link{equal7}},
\code{\link{equal8}},
\code{\link{usclass1}},
\code{\link{usclass2}},
\code{\link{usclass3}},
\code{\link{usclass4}},
\code{\link{usclass6}},
\code{\link{usclass7}},
\code{\link{usclass8}}
}
\concept{Core}
\keyword{variable}
