% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_524.R
\name{usdoenuf}
\alias{usdoenuf}
\title{U.s. protect environment}
\source{
General Social Survey https://gss.norc.org
}
\description{
usdoenuf
}
\details{
Question C. Some countries are doing more to protect the world environment than other countries are. in general, do you think that America is doing...
}
\section{Values}{

\itemize{
\item \code{1} more than enough
\item \code{2} about the right amount, or
\item \code{3} too little?
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2010 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4768/vshow}{GSS Data Explorer page for \code{usdoenuf}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{chemfree}},
\code{\link{comtype}},
\code{\link{drivless}},
\code{\link{grndemo}},
\code{\link{grngroup}},
\code{\link{grnmoney}},
\code{\link{grnsign}},
\code{\link{nomeat}},
\code{\link{recycle}}

Other Environment: 
\code{\link{chemfree}},
\code{\link{drivless}},
\code{\link{grndemo}},
\code{\link{grngroup}},
\code{\link{grnmoney}},
\code{\link{grnsign}},
\code{\link{nomeat}},
\code{\link{recycle}}
}
\concept{Environment}
\concept{ISSP Environment}
\concept{United States}
\keyword{variable}
