% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_243.R
\name{usedmntl}
\alias{usedmntl}
\title{Have sought care with mental health since plan}
\source{
General Social Survey https://gss.norc.org
}
\description{
usedmntl
}
\details{
Question Since you joined this plan, have you sought medical care for problems with your emotions, nerves, or mental health?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2189/vshow}{GSS Data Explorer page for \code{usedmntl}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1998 \tab 1639 \tab 148 \tab 1039 \tab 6 \tab 2832 \cr
   Total \tab 1639 \tab 148 \tab 1039 \tab 6 \tab 2832 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{anydoc}},
\code{\link{doclist}},
\code{\link{hlthinsr}},
\code{\link{mntlcare}},
\code{\link{psycdrct}},
\code{\link{psycfdoc}},
\code{\link{psyckid3}},
\code{\link{psycpsyc}},
\code{\link{swithlth}}
}
\concept{Mental Health}
\keyword{variable}
