% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_461.R
\name{usefrds}
\alias{usefrds}
\title{Ok to have frds just bc they are of use to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
usefrds
}
\details{
Question Please select a response to indicate how much you agree or disagree with each of the following statements.
D. It is all right to develop friendships with people just because you know they can be of use to you
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4232/vshow}{GSS Data Explorer page for \code{usefrds}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab Total \cr
   2002 \tab 1616 \tab 24 \tab 109 \tab 211 \tab 475 \tab 308 \tab 13 \tab 9 \tab 2765 \cr
   Total \tab 1616 \tab 24 \tab 109 \tab 211 \tab 475 \tab 308 \tab 13 \tab 9 \tab 2765 \cr
}
}

\seealso{
Other ISSP Social Networks & Support Systems: 
\code{\link{advantge}},
\code{\link{aidkids}},
\code{\link{comyear}},
\code{\link{demands}},
\code{\link{helpfrds}},
\code{\link{localgvt}},
\code{\link{othfrd}},
\code{\link{trustpeo}},
\code{\link{wantbest}}
}
\concept{ISSP Social Networks & Support Systems}
\keyword{variable}
