% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_280.R
\name{usehome}
\alias{usehome}
\title{Does r use computer at home}
\source{
General Social Survey https://gss.norc.org
}
\description{
usehome
}
\details{
Question B. Do you, yourself, ever use a computer at home?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2550/vshow}{GSS Data Explorer page for \code{usehome}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2403 \tab 376 \tab 15 \tab 23 \tab 2817 \cr
   Total \tab 2403 \tab 376 \tab 15 \tab 23 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{comphome}},
\code{\link{ecomhrh}},
\code{\link{ecomminh}},
\code{\link{numhome}},
\code{\link{paywww}},
\code{\link{perminh}},
\code{\link{wwwhome}},
\code{\link{wwwline1}},
\code{\link{wwwline2}}
}
\concept{Information Society}
\keyword{variable}
