% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_253.R
\name{uselang}
\alias{uselang}
\title{How often does r use language other than english}
\source{
General Social Survey https://gss.norc.org
}
\description{
uselang
}
\details{
Question F. How often do you use that language in your everyday life? (READ CATEGORIES)
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} less than once a week
\item \code{3} once a week
\item \code{4} several times a week
\item \code{5} once a day
\item \code{6} several times a day
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2288/vshow}{GSS Data Explorer page for \code{uselang}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab never \tab less than once a week \tab once a week \tab several times a week \tab once a day \tab several times a day \tab no answer \tab Total \cr
   2000 \tab 2450 \tab 79 \tab 73 \tab 28 \tab 34 \tab 19 \tab 126 \tab 8 \tab 2817 \cr
   Total \tab 2450 \tab 79 \tab 73 \tab 28 \tab 34 \tab 19 \tab 126 \tab 8 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{contjpn}},
\code{\link{contmslm}},
\code{\link{getlang}},
\code{\link{othbest}},
\code{\link{othlang}},
\code{\link{othlang1}},
\code{\link{othlang2}},
\code{\link{parlang}},
\code{\link{spklang}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
