% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_191.R
\name{usenthus}
\alias{usenthus}
\title{Thinking us how often feel enthusiastic?}
\source{
General Social Survey https://gss.norc.org
}
\description{
usenthus
}
\details{
Question Now I want to ask some questions about how you feel when you think about this country, the United States. When you think about the United States:
D. When you think about the United States, how often do you feel ENTHUSIASTIC? Do you feel enthusiastic:
}
\section{Values}{

\itemize{
\item \code{1} always
\item \code{2} most of the time
\item \code{3} some of the time
\item \code{4} rarely
\item \code{5} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1792/vshow}{GSS Data Explorer page for \code{usenthus}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab always \tab most of the time \tab some of the time \tab rarely \tab never \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2199 \tab 76 \tab 185 \tab 292 \tab 108 \tab 31 \tab 11 \tab 2 \tab 2904 \cr
   Total \tab 2199 \tab 76 \tab 185 \tab 292 \tab 108 \tab 31 \tab 11 \tab 2 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{usangry}},
\code{\link{ushopefl}},
\code{\link{usupset}}
}
\concept{Emotions}
\keyword{variable}
