% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_504.R
\name{useskill}
\alias{useskill}
\title{How much past skills can you make use in present}
\source{
General Social Survey https://gss.norc.org
}
\description{
useskill
}
\details{
Question How much of your past work experience and/or job skills can you make use of in your present job?
}
\section{Values}{

\itemize{
\item \code{1} almost none
\item \code{2} a little
\item \code{3} a lot
\item \code{4} almost all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   B/C/- \tab 2016, 2018 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4582/vshow}{GSS Data Explorer page for \code{useskill}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab almost none \tab a little \tab a lot \tab almost all \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1967 \tab 89 \tab 160 \tab 195 \tab 361 \tab 30 \tab 30 \tab 2832 \cr
   2006 \tab 3492 \tab 101 \tab 230 \tab 279 \tab 401 \tab 5 \tab 2 \tab 4510 \cr
   2016 \tab 1927 \tab 89 \tab 181 \tab 242 \tab 418 \tab 3 \tab 7 \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   Total \tab 9734 \tab 279 \tab 571 \tab 716 \tab 1180 \tab 38 \tab 39 \tab 12557 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{absent}},
\code{\link{chngwork}},
\code{\link{edcskill}},
\code{\link{helporg1}},
\code{\link{jobskill}},
\code{\link{prideorg}},
\code{\link{proudwrk}},
\code{\link{stayorg3}},
\code{\link{wktenure}}
}
\concept{Work Orientation}
\keyword{variable}
