% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_504.R
\name{useskill}
\alias{useskill}
\title{How much past skills can you make use in present}
\source{
General Social Survey https://gss.norc.org
}
\description{
useskill
}
\details{
Question How much of your past work experience and/or job skills can you make use of in your present job?
}
\section{Values}{

\itemize{
\item \code{1} almost none
\item \code{2} a little
\item \code{3} a lot
\item \code{4} almost all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab B/C/- \tab partial \cr
   2018 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4582/vshow}{GSS Data Explorer page for \code{useskill}}.
}

\seealso{
Other Work Orientation: 
\code{\link{absent}},
\code{\link{chngwork}},
\code{\link{edcskill}},
\code{\link{helporg1}},
\code{\link{jobskill}},
\code{\link{prideorg}},
\code{\link{proudwrk}},
\code{\link{stayorg3}},
\code{\link{wktenure}}

Other Job: 
\code{\link{absent}},
\code{\link{edcskill}},
\code{\link{jobskill}},
\code{\link{wktenure}}

Other Work: 
\code{\link{absent}},
\code{\link{chngwork}},
\code{\link{edcskill}},
\code{\link{helporg1}},
\code{\link{jobskill}},
\code{\link{prideorg}},
\code{\link{proudwrk}},
\code{\link{stayorg3}},
\code{\link{wktenure}}
}
\concept{Job}
\concept{Work}
\concept{Work Orientation}
\keyword{variable}
