% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_599.R
\name{usintl}
\alias{usintl}
\title{Take active part in world affairs}
\source{
General Social Survey https://gss.norc.org
}
\description{
usintl
}
\details{
Question Do you think it will be best for the future of this country if we take an active part in world affairs, or if we stay out of world affairs?
}
\section{Values}{

\itemize{
\item \code{1} active part
\item \code{2} stay out
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/- \tab full \cr
   1975 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1978 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1988 \tab A/B \tab partial \cr
   1989 \tab A/B \tab partial \cr
   1990 \tab A/B \tab partial \cr
   1991 \tab A/B \tab partial \cr
   1993 \tab A/B \tab partial \cr
   1994 \tab A/B \tab partial \cr
   2006 \tab D/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5486/vshow}{GSS Data Explorer page for \code{usintl}}.
}

\seealso{
Other War and International: 
\code{\link{commun}},
\code{\link{commun10}},
\code{\link{russia}},
\code{\link{usun}},
\code{\link{uswar}},
\code{\link{uswary}}
}
\concept{War and International}
\keyword{variable}
