% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_599.R
\name{usintl}
\alias{usintl}
\title{Take active part in world affairs}
\source{
General Social Survey https://gss.norc.org
}
\description{
usintl
}
\details{
Question Do you think it will be best for the future of this country if we take an active part in world affairs, or if we stay out of world affairs?
}
\section{Values}{

\itemize{
\item \code{1} active part
\item \code{2} stay out
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab -/- \tab full \cr
   1975 \tab -/- \tab full \cr
   1976 \tab -/- \tab full \cr
   1978 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1984 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1988 \tab A/B \tab partial \cr
   1989 \tab A/B \tab partial \cr
   1990 \tab A/B \tab partial \cr
   1991 \tab A/B \tab partial \cr
   1993 \tab A/B \tab partial \cr
   1994 \tab A/B \tab partial \cr
   2006 \tab D/- \tab partial \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5486/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab active part \tab don't know \tab no answer \tab stay out \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab - \tab 985 \tab 52 \tab 9 \tab 458 \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab - \tab 900 \tab 56 \tab 6 \tab 528 \tab - \tab 1490 \cr
   1976 \tab - \tab 945 \tab 68 \tab 3 \tab 483 \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 979 \tab 61 \tab 2 \tab 490 \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 1077 \tab 86 \tab 8 \tab 689 \tab - \tab 1860 \cr
   1983 \tab - \tab 1038 \tab 66 \tab 7 \tab 488 \tab - \tab 1599 \cr
   1984 \tab - \tab 942 \tab 88 \tab 24 \tab 419 \tab - \tab 1473 \cr
   1985 \tab 783 \tab 525 \tab 18 \tab 6 \tab 202 \tab - \tab 1534 \cr
   1986 \tab - \tab 931 \tab 52 \tab 27 \tab 460 \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 493 \tab 636 \tab 36 \tab 6 \tab 310 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 687 \tab 36 \tab 2 \tab 281 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 633 \tab 38 \tab 4 \tab 253 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 752 \tab 25 \tab - \tab 247 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 725 \tab 49 \tab - \tab 306 \tab - \tab 1606 \cr
   1994 \tab 2485 \tab 324 \tab 18 \tab 6 \tab 159 \tab - \tab 2992 \cr
   1996 \tab 2904 \tab - \tab - \tab - \tab - \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2992 \tab 887 \tab 54 \tab 4 \tab 573 \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab - \tab - \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab - \tab - \tab - \tab - \tab - \tab 3309 \tab 3309 \cr
   Total \tab 48617 \tab 12966 \tab 803 \tab 114 \tab 6346 \tab 6853 \tab 75699 \cr
}
}

\seealso{
Other War and International: 
\code{\link{commun}},
\code{\link{commun10}},
\code{\link{russia}},
\code{\link{usun}},
\code{\link{uswar}},
\code{\link{uswary}}

Other Foreign Affairs: 
\code{\link{usun}},
\code{\link{uswar}},
\code{\link{uswary}}

Other United States: 
\code{\link{usun}},
\code{\link{uswar}},
\code{\link{uswary}}
}
\concept{Foreign Affairs}
\concept{United States}
\concept{War and International}
\keyword{variable}
