% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_680.R
\name{usmca1}
\alias{usmca1}
\title{R's familiarity with usmca}
\source{
General Social Survey https://gss.norc.org
}
\description{
usmca1
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} a lot
\item \code{2} quite a bit
\item \code{3} not much
\item \code{4} nothing at all
\item \code{NA(d)} can't choose
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
