% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_680.R
\name{usmca1}
\alias{usmca1}
\title{R's familiarity with usmca}
\source{
General Social Survey https://gss.norc.org
}
\description{
usmca1
}
\details{
Question How much have you heard or read about the United States-Mexico-Canada Agreement (USMCA)?
}
\section{Values}{

\itemize{
\item \code{1} a lot
\item \code{2} quite a bit
\item \code{3} not much
\item \code{4} nothing at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8185/vshow}{GSS Data Explorer page for \code{usmca1}}.
}

\seealso{
Other ISSP Citizenship: 
\code{\link{leftrght1}},
\code{\link{poltalk}},
\code{\link{servpub}}
}
\concept{ISSP Citizenship}
\keyword{variable}
