% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_680.R
\name{usmca2}
\alias{usmca2}
\title{Does the us benefit from usmca}
\source{
General Social Survey https://gss.norc.org
}
\description{
usmca2
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} greatly benefits
\item \code{2} largely benefits
\item \code{3} somewhat benefits
\item \code{4} benefits only a little
\item \code{5} does not benefit at all
\item \code{NA(d)} can't choose
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
