% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_681.R
\name{usmca2}
\alias{usmca2}
\title{Does the us benefit from usmca}
\source{
General Social Survey https://gss.norc.org
}
\description{
usmca2
}
\details{
Question Generally speaking, would you say that America benefits or does not benefit from being a member of United States-Mexico-Canada Agreement (USMCA)?
}
\section{Values}{

\itemize{
\item \code{1} greatly benefits
\item \code{2} largely benefits
\item \code{3} somewhat benefits
\item \code{4} benefits only a little
\item \code{5} does not benefit at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8186/vshow}{GSS Data Explorer page for \code{usmca2}}.
}

\concept{ISSP Citizenship}
\keyword{variable}
