% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_256.R
\name{usmixed}
\alias{usmixed}
\title{What percentage of population is of mixed heritage}
\source{
General Social Survey https://gss.norc.org
}
\description{
usmixed
}
\details{
Question People may come from one racial group such as being White, Black/African-American, Hispanic, Asian, or American Indian, or people may come from two or more of these groups. What percent of people in the United States today do you think have parents or their ancestors from two or more racial groups?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2316/vshow}{GSS Data Explorer page for \code{usmixed}}.
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{blkchng}},
\code{\link{hispchng}},
\code{\link{jewschng}},
\code{\link{usamind}},
\code{\link{usasn}},
\code{\link{usblk}},
\code{\link{ushisp}},
\code{\link{usjews}},
\code{\link{whtchng}}

Other United States: 
\code{\link{blkchng}},
\code{\link{hispchng}},
\code{\link{jewschng}},
\code{\link{usamind}},
\code{\link{usasn}},
\code{\link{usblk}},
\code{\link{ushisp}},
\code{\link{usjews}},
\code{\link{whtchng}}
}
\concept{Multi-Ethnic United States}
\concept{United States}
\keyword{variable}
