% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_599.R
\name{usun}
\alias{usun}
\title{Remain in u.n. or pull out}
\source{
General Social Survey https://gss.norc.org
}
\description{
usun
}
\details{
Question Do you think our government should continue to belong to the United Nations, or should we pull out of it now?
}
\section{Values}{

\itemize{
\item \code{1} stay in
\item \code{2} pull out
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   -/- \tab 1973, 1975, 1976, 1978, 1982, 1983, 1985, 1986 \cr
   A/B \tab 1988, 1989, 1990, 1991, 1993, 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5487/vshow}{GSS Data Explorer page for \code{usun}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab stay in \tab pull out \tab don't know \tab no answer \tab Total \cr
   1973 \tab - \tab 1188 \tab 224 \tab 84 \tab 8 \tab 1504 \cr
   1975 \tab - \tab 1120 \tab 270 \tab 98 \tab 2 \tab 1490 \cr
   1976 \tab - \tab 1088 \tab 286 \tab 123 \tab 2 \tab 1499 \cr
   1978 \tab - \tab 1216 \tab 197 \tab 116 \tab 3 \tab 1532 \cr
   1982 \tab - \tab 1408 \tab 317 \tab 131 \tab 4 \tab 1860 \cr
   1983 \tab - \tab 1278 \tab 220 \tab 98 \tab 3 \tab 1599 \cr
   1985 \tab - \tab 1201 \tab 260 \tab 68 \tab 5 \tab 1534 \cr
   1986 \tab - \tab 1124 \tab 239 \tab 84 \tab 23 \tab 1470 \cr
   1988 \tab 493 \tab 771 \tab 155 \tab 60 \tab 2 \tab 1481 \cr
   1989 \tab 531 \tab 829 \tab 121 \tab 56 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 772 \tab 100 \tab 54 \tab 2 \tab 1372 \cr
   1991 \tab 493 \tab 897 \tab 95 \tab 31 \tab 1 \tab 1517 \cr
   1993 \tab 526 \tab 920 \tab 106 \tab 53 \tab 1 \tab 1606 \cr
   1994 \tab 2485 \tab 415 \tab 65 \tab 25 \tab 2 \tab 2992 \cr
   Total \tab 4972 \tab 14227 \tab 2655 \tab 1081 \tab 58 \tab 22993 \cr
}
}

\seealso{
Other War and International: 
\code{\link{commun}},
\code{\link{commun10}},
\code{\link{russia}},
\code{\link{usintl}},
\code{\link{uswar}},
\code{\link{uswary}}
}
\concept{War and International}
\keyword{variable}
