% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_599.R
\name{uswar}
\alias{uswar}
\title{Expect u.s. in war within 10 years}
\source{
General Social Survey https://gss.norc.org
}
\description{
uswar
}
\details{
Question Do you expect the United States to fight in another war within the next ten years?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1973 \tab - \tab full \cr
   1975 \tab - \tab full \cr
   1976 \tab - \tab full \cr
   1978 \tab - \tab full \cr
   1982 \tab - \tab full \cr
   1983 \tab - \tab full \cr
   1985 \tab - \tab full \cr
   2010 \tab B \tab partial \cr
   2012 \tab B \tab partial \cr
   2014 \tab B \tab partial \cr
   2016 \tab B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5484/vshow}{GSS Data Explorer page for \code{uswar}}.
}

\seealso{
Other War and International: 
\code{\link{brazil}},
\code{\link{canada}},
\code{\link{china}},
\code{\link{commun}},
\code{\link{commun10}},
\code{\link{egypt}},
\code{\link{england}},
\code{\link{israel}},
\code{\link{japan}},
\code{\link{russia}},
\code{\link{usintl}},
\code{\link{usun}},
\code{\link{uswary}}
}
\concept{War and International}
\keyword{variable}
