% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_599.R
\name{uswary}
\alias{uswary}
\title{Expect u.s. in world war in 10 years}
\source{
General Social Survey https://gss.norc.org
}
\description{
uswary
}
\details{
Question Do you expect the United States to fight in another world war within the next ten years?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1976 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1988 \tab A/B \tab partial \cr
   1989 \tab A/B \tab partial \cr
   1990 \tab A/B \tab partial \cr
   1991 \tab A/B \tab partial \cr
   1993 \tab A/B \tab partial \cr
   1994 \tab A/B \tab partial \cr
   1996 \tab A/B \tab partial \cr
   1998 \tab A/B \tab partial \cr
   2000 \tab A/B \tab partial \cr
   2002 \tab A/B \tab partial \cr
   2004 \tab A/B \tab partial \cr
   2006 \tab A/B \tab partial \cr
   2008 \tab A/B \tab partial \cr
   2010 \tab A/- \tab partial \cr
   2012 \tab A/- \tab partial \cr
   2014 \tab A/- \tab partial \cr
   2016 \tab A/- \tab partial \cr
   2018 \tab A/- \tab partial \cr
   2021 \tab A/- \tab partial \cr
   2022 \tab A/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5485/vshow}{GSS Data Explorer page for \code{uswary}}.
}

\seealso{
Other War and International: 
\code{\link{commun}},
\code{\link{commun10}},
\code{\link{russia}},
\code{\link{usintl}},
\code{\link{usun}},
\code{\link{uswar}}

Other Foreign Affairs: 
\code{\link{usintl}},
\code{\link{usun}},
\code{\link{uswar}}

Other Split Ballots: 
\code{\link{russia}},
\code{\link{uswar}}

Other United States: 
\code{\link{usintl}},
\code{\link{usun}},
\code{\link{uswar}}
}
\concept{Foreign Affairs}
\concept{Split Ballots}
\concept{United States}
\concept{War and International}
\keyword{variable}
