% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_637.R
\name{uswaryv}
\alias{uswaryv}
\title{Expect u.s. in war within 10 years (with volunteered response on web)}
\source{
General Social Survey https://gss.norc.org
}
\description{
uswaryv
}
\details{
Question Do you expect the United States to fight in another world war within the next ten years?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{3} undecided (phone mode only: volunteered)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2021 \tab A \tab partial \cr
   2022 \tab A \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7584/vshow}{GSS Data Explorer page for \code{uswaryv}}.
}

\keyword{variable}
