% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_190.R
\name{usworry}
\alias{usworry}
\title{How often do you feel worried for us?}
\source{
General Social Survey https://gss.norc.org
}
\description{
usworry
}
\details{
Question Now I want to ask some questions about how you feel when you think about this country, the United States. When you think about the United States:
A. How often do you feel WORRIED? Do you feel worried:
}
\section{Values}{

\itemize{
\item \code{1} always
\item \code{2} most of the time
\item \code{3} some of the time
\item \code{4} rarely
\item \code{5} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1789/vshow}{GSS Data Explorer page for \code{usworry}}.
}

\seealso{
Other Emotions: 
\code{\link{amimp}},
\code{\link{chnang16}},
\code{\link{likeoth}},
\code{\link{notam}},
\code{\link{othresp}},
\code{\link{reactok}},
\code{\link{selfresp}},
\code{\link{usfrustr}},
\code{\link{ussat}}

Other United States: 
\code{\link{amimp}},
\code{\link{notam}},
\code{\link{usfrustr}},
\code{\link{ussat}}
}
\concept{Emotions}
\concept{United States}
\keyword{variable}
