% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_207.R
\name{valhlth}
\alias{valhlth}
\title{Estimated value contributed in health}
\source{
General Social Survey https://gss.norc.org
}
\description{
valhlth
}
\details{
Question Of the total amount you gave to (AREA IN TOTHLTH-TOTOTH what is the estimated dollar value of contributed property, e.g., clothing, furniture, real estate, equipment, etc.? (REPEAT FOR EACH ITEM MENTIONED IN TOTHLTH-TOTOTH)
RECORD above UNDER VALHLTH-VALOTH. (ESTIMATED CASH-VALUE OF PROPERTY CONTRIBUTIONS ONLY.)
A. Health
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1960/vshow}{GSS Data Explorer page for \code{valhlth}}.

Counts by year:\tabular{lllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 10 \tab 100 \tab 1000 \tab 12 \tab 120 \tab 1200 \tab 125 \tab 15 \tab 150 \tab 18 \tab 20 \tab 200 \tab 25 \tab 300 \tab 40 \tab 400 \tab 50 \tab 75 \tab 900 \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2653 \tab 122 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 9 \tab 93 \tab 2904 \cr
   Total \tab 2653 \tab 122 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 9 \tab 93 \tab 2904 \cr
}
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{totart}},
\code{\link{totfound}},
\code{\link{totinfrm}},
\code{\link{totintl}},
\code{\link{tototh}},
\code{\link{totpol}},
\code{\link{totwork}},
\code{\link{totyouth}},
\code{\link{valeduc}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\keyword{variable}
