% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_208.R
\name{valpol}
\alias{valpol}
\title{Estimated value contributed in political org. etc}
\source{
General Social Survey https://gss.norc.org
}
\description{
valpol
}
\details{
Question Of the total amount you gave to (AREA IN TOTHLTH-TOTOTH what is the estimated dollar value of contributed property, e.g., clothing, furniture, real estate, equipment, etc.? (REPEAT FOR EACH ITEM MENTIONED IN TOTHLTH-TOTOTH)
RECORD above UNDER VALHLTH-VALOTH. (ESTIMATED CASH-VALUE OF PROPERTY CONTRIBUTIONS ONLY.)
J. Political organizations or campaigns
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1969/vshow}{GSS Data Explorer page for \code{valpol}}.
}

\seealso{
Other Markets & Exchange - Giving & Volunteering: 
\code{\link{valart}},
\code{\link{valenvir}},
\code{\link{valfound}},
\code{\link{valhuman}},
\code{\link{valpub}},
\code{\link{valrec}},
\code{\link{valrelig}},
\code{\link{valwork}},
\code{\link{valyouth}}

Other Voluntary Associations: 
\code{\link{valart}},
\code{\link{valenvir}},
\code{\link{valfound}},
\code{\link{valhuman}},
\code{\link{valpub}},
\code{\link{valrec}},
\code{\link{valrelig}},
\code{\link{valwork}},
\code{\link{valyouth}}
}
\concept{Markets & Exchange - Giving & Volunteering}
\concept{Voluntary Associations}
\keyword{variable}
