% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_662.R
\name{vaxhstncy}
\alias{vaxhstncy}
\title{Parents choose not to vaccinate their children}
\source{
General Social Survey https://gss.norc.org
}
\description{
vaxhstncy
}
\details{
Question Now, we have a few questions for you about different kinds of vaccinations. Please tell us the extent to which you agree or disagree with the following statement: Parents should be able to refuse to vaccinate their children against measles, mumps, and r

NOTE: This variable originates on the GSS Next follow-on, and should not be analyzed without the GSS Next-specific weights WTSSPS_NEXT or WTSSNRPS_NEXT
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7948/vshow}{GSS Data Explorer page for \code{vaxhstncy}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab skipped on web \tab Total \cr
   2022 \tab 2309 \tab 98 \tab 203 \tab 282 \tab 371 \tab 280 \tab 1 \tab 3544 \cr
   Total \tab 2309 \tab 98 \tab 203 \tab 282 \tab 371 \tab 280 \tab 1 \tab 3544 \cr
}
}

\seealso{
Other GSS Next: 
\code{\link{biokids}},
\code{\link{covemply}},
\code{\link{covid12}},
\code{\link{fluvax}},
\code{\link{malekids}},
\code{\link{pandinc}},
\code{\link{pandmet}},
\code{\link{vaxkids}},
\code{\link{vaxsafe}}
}
\concept{GSS Next}
\keyword{variable}
