% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_436.R
\name{verdict}
\alias{verdict}
\title{Worse type of judicial mistake}
\source{
General Social Survey https://gss.norc.org
}
\description{
verdict
}
\details{
Question All systems of justice make mistakes, but which do you think is worse?
To convict an innocent person
or
To let a guilty person go free
}
\section{Values}{

\itemize{
\item \code{1} convict an innocent person
\item \code{2} let a guilty person go free
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2016 \cr
   A/B/C \tab 1990, 1996 \cr
   D/-/- \tab 2006 \cr
   Full \tab 1985 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4002/vshow}{GSS Data Explorer page for \code{verdict}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab convict an innocent person \tab let a guilty person go free \tab can't choose \tab no answer \tab Total \cr
   1985 \tab 857 \tab 419 \tab 138 \tab 109 \tab 11 \tab 1534 \cr
   1990 \tab 155 \tab 684 \tab 243 \tab 249 \tab 41 \tab 1372 \cr
   1996 \tab 1572 \tab 805 \tab 300 \tab 198 \tab 29 \tab 2904 \cr
   2006 \tab 2992 \tab 1030 \tab 428 \tab 52 \tab 8 \tab 4510 \cr
   2016 \tab 1477 \tab 1036 \tab 291 \tab 49 \tab 14 \tab 2867 \cr
   Total \tab 7053 \tab 3974 \tab 1400 \tab 657 \tab 103 \tab 13187 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{databank}},
\code{\link{eqincome}},
\code{\link{manhold}},
\code{\link{manread}},
\code{\link{mantail}},
\code{\link{mantap}},
\code{\link{opprof}},
\code{\link{oprich}},
\code{\link{progtax}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
