% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_91.R
\name{vetaid}
\alias{vetaid}
\title{Any in hh receive mil or vet benefits}
\source{
General Social Survey https://gss.norc.org
}
\description{
vetaid
}
\details{
Question Are you or any members of your present household currently receiving any pay or benefits from either the military or the Veterans Administration?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/828/vshow}{GSS Data Explorer page for \code{vetaid}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1984 \tab - \tab 145 \tab 1298 \tab 2 \tab 28 \tab 1473 \cr
   Total \tab 0 \tab 145 \tab 1298 \tab 2 \tab 28 \tab 1473 \cr
}
}

\seealso{
Other Core: 
\code{\link{defwrkev}},
\code{\link{guerilla}},
\code{\link{jobtrain}},
\code{\link{landwar}},
\code{\link{lessnuke}},
\code{\link{morenuke}},
\code{\link{nonuke}},
\code{\link{nukewar}},
\code{\link{upgrade}()}
}
\concept{Core}
\keyword{variable}
