% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_577.R
\name{vetkind}
\alias{vetkind}
\title{Branch of service}
\source{
General Social Survey https://gss.norc.org
}
\description{
vetkind
}
\details{
Question Have you ever been on active duty for military training or service for two consecutive months or more?
IF YES: What was your total time on active duty?
A. In what branch of the service was that?
}
\section{Values}{

\itemize{
\item \code{0} several
\item \code{1} air force guard
\item \code{2} air force
\item \code{3} navy
\item \code{4} army
\item \code{5} national guard
\item \code{6} marines
\item \code{7} coast guard
\item \code{8} public health
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1974, 1975, 1977, 1978, 1982, 1983, 1984 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5286/vshow}{GSS Data Explorer page for \code{vetkind}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab several \tab air force guard \tab air force \tab navy \tab army \tab national guard \tab marines \tab coast guard \tab no answer \tab public health \tab Total \cr
   1974 \tab 1142 \tab 8 \tab 2 \tab 50 \tab 72 \tab 174 \tab 8 \tab 16 \tab 4 \tab 8 \tab - \tab 1484 \cr
   1975 \tab 1162 \tab 3 \tab - \tab 39 \tab 63 \tab 191 \tab 6 \tab 18 \tab 4 \tab 4 \tab - \tab 1490 \cr
   1977 \tab 1188 \tab 3 \tab 1 \tab 49 \tab 57 \tab 195 \tab 9 \tab 16 \tab 4 \tab 8 \tab - \tab 1530 \cr
   1978 \tab 1205 \tab 3 \tab 1 \tab 56 \tab 70 \tab 166 \tab 11 \tab 14 \tab 2 \tab 4 \tab - \tab 1532 \cr
   1982 \tab 1543 \tab 1 \tab 1 \tab 46 \tab 52 \tab 174 \tab 9 \tab 22 \tab 3 \tab 8 \tab 1 \tab 1860 \cr
   1983 \tab 1301 \tab 3 \tab 1 \tab 51 \tab 70 \tab 137 \tab 5 \tab 23 \tab 5 \tab 3 \tab - \tab 1599 \cr
   1984 \tab 1205 \tab 4 \tab - \tab 40 \tab 48 \tab 118 \tab 5 \tab 25 \tab 2 \tab 26 \tab - \tab 1473 \cr
   Total \tab 8746 \tab 25 \tab 6 \tab 331 \tab 432 \tab 1155 \tab 53 \tab 134 \tab 24 \tab 61 \tab 1 \tab 10968 \cr
}
}

\seealso{
Other Military Service & Work: 
\code{\link{spdays}},
\code{\link{sphour}},
\code{\link{vetyears}},
\code{\link{workdy}},
\code{\link{workhr}}
}
\concept{Military Service & Work}
\keyword{variable}
