% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_577.R
\name{vetyears}
\alias{vetyears}
\title{Years in armed forces}
\source{
General Social Survey https://gss.norc.org
}
\description{
vetyears
}
\details{
Question Have you ever been on active duty for military training or service for two consecutive months or more?
IF YES: What was your total time on active duty?
}
\section{Values}{

\itemize{
\item \code{0} no active duty
\item \code{1} yes, less than 2 years
\item \code{2} yes, 2-4 years
\item \code{3} yes, more than 4 years
\item \code{4} some, don't know how long
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1974 \tab -/-/- \tab full \cr
   1975 \tab -/-/- \tab full \cr
   1977 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1983 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1985 \tab -/-/- \tab full \cr
   1988 \tab A/B/- \tab partial \cr
   1989 \tab A/B/- \tab partial \cr
   1990 \tab A/B/- \tab partial \cr
   1991 \tab A/B/- \tab partial \cr
   1993 \tab A/B/- \tab partial \cr
   1994 \tab A/B/- \tab partial \cr
   2010 \tab A/B/C \tab full \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2021 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5285/vshow}{GSS Data Explorer page for \code{vetyears}}.
}

\seealso{
Other Military Service & Work: 
\code{\link{spdays}},
\code{\link{sphour}},
\code{\link{vetkind}},
\code{\link{workdy}},
\code{\link{workhr}}

Other Veteran Status: 
\code{\link{vetkind}}
}
\concept{Military}
\concept{Military Service & Work}
\concept{Veteran Status}
\keyword{variable}
