% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_392.R
\name{viglab06}
\alias{viglab06}
\title{X is experiencing alchl dependencedepression schizophrenia}
\source{
General Social Survey https://gss.norc.org
}
\description{
viglab06
}
\details{
Question In your opinion, how likely is it that (NAME) is experiencing…
IF VIGVER06 = 1, DISPLAY Alcohol dependence?
IF VIGVER06 = 2, DISPLAY A major depression?
IF VIGVER06 = 3, DISPLAY Schizophrenia?
IF VIGVER06 = 4, Skip to MHTRTOTH
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3564/vshow}{GSS Data Explorer page for \code{viglab06}}.
}

\seealso{
Other Mental Health III: 
\code{\link{mhothrel}},
\code{\link{mhothyou}},
\code{\link{mhtrtoth}},
\code{\link{seemhpub}}

Other Mental Health: 
\code{\link{mhothrel}},
\code{\link{mhothyou}},
\code{\link{mhtrtoth}},
\code{\link{seemhpub}}
}
\concept{Mental Health}
\concept{Mental Health III}
\keyword{variable}
