% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_211.R
\name{viglabel}
\alias{viglabel}
\title{How likely x is experiencing vignettes}
\source{
General Social Survey https://gss.norc.org
}
\description{
viglabel
}
\details{
Question In your opinion, how likely is it that NAME is experiencing (READ A-E) -- very likely, somewhat likely, not very likely, or not at all likely? (IN PART E, READ ONLY THE PHRASE FOLLOWING THE LETTER OF THE VIGNETTE)
E. REFER TO VIGNETTE ID, CHECK BELOW AND READ ONLY 1 PHRASE OF THE FOLLOWING:
Vignette A: alcohol dependence
Vignette B: a major depression
Vignette C: schizophrenia
Vignette D: a drug problem
Vignette E: GO TO DECTREAT
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} not at all likely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2018 \tab B/C/- \tab partial \cr
   2024 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1998/vshow}{GSS Data Explorer page for \code{viglabel}}.
}

\seealso{
Other Mental Health: 
\code{\link{breakdwn}},
\code{\link{decmoney}},
\code{\link{dectreat}},
\code{\link{genetics}},
\code{\link{godswill}},
\code{\link{imprvown}},
\code{\link{mentlill}},
\code{\link{physill}},
\code{\link{upsdowns}}
}
\concept{Mental Health}
\keyword{variable}
