% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_212.R
\name{vigsoc}
\alias{vigsoc}
\title{How willing r would spend time socializing with x}
\source{
General Social Survey https://gss.norc.org
}
\description{
vigsoc
}
\details{
Question How willing would you be (READ A-F) -- definitely willing, probably willing, probably unwilling, or definitely unwilling?
B. To spend an evening socializing with NAME?
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely willing
\item \verb{[2]} probably willing
\item \verb{[3]} probably unwilling
\item \verb{[4]} definitely unwilling
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2004/vshow}{GSS Data Explorer page for \code{vigsoc}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely willing \tab probably willing \tab probably unwilling \tab definitely unwilling \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 212 \tab 540 \tab 344 \tab 295 \tab 31 \tab 22 \tab 2904 \cr
   2006 \tab 3073 \tab 240 \tab 609 \tab 362 \tab 173 \tab 38 \tab 15 \tab 4510 \cr
   2018 \tab 1175 \tab 201 \tab 508 \tab 266 \tab 151 \tab 16 \tab 31 \tab 2348 \cr
   Total \tab 5708 \tab 653 \tab 1657 \tab 972 \tab 619 \tab 85 \tab 68 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{hurtoth}},
\code{\link{hurtself}},
\code{\link{imprvtrt}},
\code{\link{tlkfam}},
\code{\link{vigfrnd}},
\code{\link{viggrp}},
\code{\link{vigmar}},
\code{\link{vignei}},
\code{\link{vigwork}}
}
\concept{Mental Health}
\keyword{variable}
