% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_210.R
\name{vigversn}
\alias{vigversn}
\title{Vignette version number}
\source{
General Social Survey https://gss.norc.org
}
\description{
vigversn
}
\details{
Question Next I'm going to describe a person -- lets call him/her (REFER TO NAME IN VIGNETTE). After I read a description of him/her I'll ask you some questions about how you think and feel about him/her. There are no right or wrong answers. I'm only interested in what you think of NAME.
}
\section{Values}{

\itemize{
\item \code{1} 1: alcohol dependence, john, white, eighth grade
\item \code{2} 2: alcohol dependence, john, african american, eighth grade
\item \code{3} 3: alcohol dependence, john, hispanic, eighth grade
\item \code{4} 4: alcohol dependence, john, white, high school
\item \code{5} 5: alcohol dependence, john, african american, high school
\item \code{6} 6: alcohol dependence, john, hispanic, high school
\item \code{7} 7: alcohol dependence, john, white, college
\item \code{8} 8: alcohol dependence, john, african american, college
\item \code{9} 9: alcohol dependence, john, hispanic, college
\item \code{10} 10: alcohol dependence, mary, white, eighth grade
\item \code{11} 11: alcohol dependence, mary, african american, eighth grade
\item \code{12} 12: alcohol dependence, mary, hispanic, eighth grade
\item \code{13} 13: alcohol dependence, mary, white, high school
\item \code{14} 14: alcohol dependence, mary, african american, high school
\item \code{15} 15: alcohol dependence, mary, hispanic, high school
\item \code{16} 16: alcohol dependence, mary, white, college
\item \code{17} 17: alcohol dependence, mary, african american, college
\item \code{18} 18: alcohol dependence, mary, hispanic, college
\item \code{19} 19: major depression, john, white, eighth grade
\item \code{20} 20: major depression, john, african american, eighth grade
\item \code{21} 21: major depression, john, hispanic, eighth grade
\item \code{22} 22: major depression, john, white, high school
\item \code{23} 23: major depression, john, african american, high school
\item \code{24} 24: major depression, john, hispanic, high school
\item \code{25} 25: major depression, john, white, college
\item \code{26} 26: major depression, john, african american, college
\item \code{27} 27: major depression, john, hispanic, college
\item \code{28} 28: major depression, mary, white, eighth grade
\item \code{29} 29: major depression, mary, african american, eighth grade
\item \code{30} 30: major depression, mary, hispanic, eighth grade
\item \code{31} 31: major depression, mary, white, high school
\item \code{32} 32: major depression, mary, african american, high school
\item \code{33} 33: major depression, mary, hispanic, high school
\item \code{34} 34: major depression, mary, white, college
\item \code{35} 35: major depression, mary, african american, college
\item \code{36} 36: major depression, mary, hispanic, college
\item \code{37} 37: schizophrenia, john, white, eighth grade
\item \code{38} 38: schizophrenia, john, african american, eighth grade
\item \code{39} 39: schizophrenia, john, hispanic, eighth grade
\item \code{40} 40: schizophrenia, john, white, high school
\item \code{41} 41: schizophrenia, john, african american, high school
\item \code{42} 42: schizophrenia, john, hispanic, high school
\item \code{43} 43: schizophrenia, john, white, college
\item \code{44} 44: schizophrenia, john, african american, college
\item \code{45} 45: schizophrenia, john, hispanic, college
\item \code{46} 46: schizophrenia, mary, white, eighth grade
\item \code{47} 47: schizophrenia, mary, african american, eighth grade
\item \code{48} 48: schizophrenia, mary, hispanic, eighth grade
\item \code{49} 49: schizophrenia, mary, white, high school
\item \code{50} 50: schizophrenia, mary, african american, high school
\item \code{51} 51: schizophrenia, mary, hispanic, high school
\item \code{52} 52: schizophrenia, mary, white, college
\item \code{53} 53: schizophrenia, mary, african american, college
\item \code{54} 54: schizophrenia, mary, hispanic, college
\item \code{55} 55: drug problem, john, white, eighth grade
\item \code{56} 56: drug problem, john, african american, eighth grade
\item \code{57} 57: drug problem, john, hispanic, eighth grade
\item \code{58} 58: drug problem, john, white, high school
\item \code{59} 59: drug problem, john, african american, high school
\item \code{60} 60: drug problem, john, hispanic, high school
\item \code{61} 61: drug problem, john, white, college
\item \code{62} 62: drug problem, john, african american, college
\item \code{63} 63: drug problem, john, hispanic, college
\item \code{64} 64: drug problem, mary, white, eighth grade
\item \code{65} 65: drug problem, mary, african american, eighth grade
\item \code{66} 66: drug problem, mary, hispanic, eighth grade
\item \code{67} 67: drug problem, mary, white, high school
\item \code{68} 68: drug problem, mary, african american, high school
\item \code{69} 69: drug problem, mary, hispanic, high school
\item \code{70} 70: drug problem, mary, white, college
\item \code{71} 71: drug problem, mary, african american, college
\item \code{72} 72: drug problem, mary, hispanic, college
\item \code{73} 73: no problem, john, white, eighth grade
\item \code{74} 74: no problem, john, african american, eighth grade
\item \code{75} 75: no problem, john, hispanic, eighth grade
\item \code{76} 76: no problem, john, white, high school
\item \code{77} 77: no problem, john, african american, high school
\item \code{78} 78: no problem, john, hispanic, high school
\item \code{79} 79: no problem, john, white, college
\item \code{80} 80: no problem, john, african american, college
\item \code{81} 81: no problem, john, hispanic, college
\item \code{82} 82: no problem, mary, white, eighth grade
\item \code{83} 83: no problem, mary, african american, eighth grade
\item \code{84} 84: no problem, mary, hispanic, eighth grade
\item \code{85} 85: no problem, mary, white, high school
\item \code{86} 86: no problem, mary, african american, high school
\item \code{87} 87: no problem, mary, hispanic, high school
\item \code{88} 88: no problem, mary, white, college
\item \code{89} 89: no problem, mary, african american, college
\item \code{90} 90: no problem, mary, hispanic, college
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1986/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 10: alcohol dependence, mary, white, eighth grade \tab 11: alcohol dependence, mary, african american, eighth grade \tab 12: alcohol dependence, mary, hispanic, eighth grade \tab 13: alcohol dependence, mary, white, high school \tab 14: alcohol dependence, mary, african american, high school \tab 15: alcohol dependence, mary, hispanic, high school \tab 16: alcohol dependence, mary, white, college \tab 17: alcohol dependence, mary, african american, college \tab 18: alcohol dependence, mary, hispanic, college \tab 19: major depression, john, white, eighth grade \tab 1:  alcohol dependence, john, white, eighth grade \tab 20: major depression, john, african american, eighth grade \tab 21: major depression, john, hispanic, eighth grade \tab 22: major depression, john, white, high school \tab 23: major depression, john, african american, high school \tab 24: major depression, john, hispanic, high school \tab 25: major depression, john, white, college \tab 26: major depression, john, african american, college \tab 27: major depression, john, hispanic, college \tab 28: major depression, mary, white, eighth grade \tab 29: major depression, mary, african american, eighth grade \tab 2:  alcohol dependence, john, african american, eighth grade \tab 30: major depression, mary, hispanic, eighth grade \tab 31: major depression, mary, white, high school \tab 32: major depression, mary, african american, high school \tab 33: major depression, mary, hispanic, high school \tab 34: major depression, mary, white, college \tab 35: major depression, mary, african american, college \tab 36: major depression, mary, hispanic, college \tab 37: schizophrenia, john, white, eighth grade \tab 38: schizophrenia, john, african american, eighth grade \tab 39: schizophrenia, john, hispanic, eighth grade \tab 3:  alcohol dependence, john, hispanic, eighth grade \tab 40: schizophrenia, john, white, high school \tab 41: schizophrenia, john, african american, high school \tab 42: schizophrenia, john, hispanic, high school \tab 43: schizophrenia, john, white, college \tab 44: schizophrenia, john, african american, college \tab 45: schizophrenia, john, hispanic, college \tab 46: schizophrenia, mary, white, eighth grade \tab 47: schizophrenia, mary, african american, eighth grade \tab 48: schizophrenia, mary, hispanic, eighth grade \tab 49: schizophrenia, mary, white, high school \tab 4:  alcohol dependence, john, white, high school \tab 50: schizophrenia, mary, african american, high school \tab 51: schizophrenia, mary, hispanic, high school \tab 52: schizophrenia, mary, white, college \tab 53: schizophrenia, mary, african american, college \tab 54: schizophrenia, mary, hispanic, college \tab 55: drug problem, john, white, eighth grade \tab 56: drug problem, john, african american, eighth grade \tab 57: drug problem, john, hispanic, eighth grade \tab 58: drug problem, john, white, high school \tab 59: drug problem, john, african american, high school \tab 5:  alcohol dependence, john, african american, high school \tab 60: drug problem, john, hispanic, high school \tab 61: drug problem, john, white, college \tab 62: drug problem, john, african american, college \tab 63: drug problem, john, hispanic, college \tab 64: drug problem, mary, white, eighth grade \tab 65: drug problem, mary, african american, eighth grade \tab 66: drug problem, mary, hispanic, eighth grade \tab 67: drug problem, mary, white, high school \tab 68: drug problem, mary, african american, high school \tab 69: drug problem, mary, hispanic, high school \tab 6:  alcohol dependence, john, hispanic, high school \tab 70: drug problem, mary, white, college \tab 71: drug problem, mary, african american, college \tab 72: drug problem, mary, hispanic, college \tab 73: no problem, john, white, eighth grade \tab 74: no problem, john, african american, eighth grade \tab 75: no problem, john, hispanic, eighth grade \tab 76: no problem, john, white, high school \tab 77: no problem, john, african american, high school \tab 78: no problem, john, hispanic, high school \tab 79: no problem, john, white, college \tab 7:  alcohol dependence, john, white, college \tab 80: no problem, john, african american, college \tab 81: no problem, john, hispanic, college \tab 82: no problem, mary, white, eighth grade \tab 83: no problem, mary, african american, eighth grade \tab 84: no problem, mary, hispanic, eighth grade \tab 85: no problem, mary, white, high school \tab 86: no problem, mary, african american, high school \tab 87: no problem, mary, hispanic, high school \tab 88: no problem, mary, white, college \tab 89: no problem, mary, african american, college \tab 8:  alcohol dependence, john, african american, college \tab 90: no problem, mary, hispanic, college \tab 9:  alcohol dependence, john, hispanic, college \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab 1484 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 1499 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab 1860 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 1473 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 1819 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1819 \cr
   1988 \tab 1481 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1481 \cr
   1989 \tab 1537 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1537 \cr
   1990 \tab 1372 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1372 \cr
   1991 \tab 1517 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1517 \cr
   1993 \tab 1606 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1606 \cr
   1994 \tab 2992 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2992 \cr
   1996 \tab 1460 \tab 22 \tab 18 \tab 18 \tab 14 \tab 12 \tab 11 \tab 18 \tab 20 \tab 22 \tab 14 \tab 15 \tab 10 \tab 16 \tab 21 \tab 17 \tab 19 \tab 15 \tab 15 \tab 14 \tab 23 \tab 21 \tab 11 \tab 12 \tab 19 \tab 17 \tab 8 \tab 24 \tab 16 \tab 19 \tab 19 \tab 19 \tab 16 \tab 13 \tab 17 \tab 20 \tab 20 \tab 16 \tab 12 \tab 12 \tab 19 \tab 19 \tab 21 \tab 10 \tab 17 \tab 11 \tab 18 \tab 18 \tab 17 \tab 17 \tab 15 \tab 14 \tab 8 \tab 17 \tab 18 \tab 16 \tab 18 \tab 15 \tab 15 \tab 21 \tab 17 \tab 16 \tab 21 \tab 13 \tab 17 \tab 14 \tab 17 \tab 16 \tab 19 \tab 18 \tab 9 \tab 13 \tab 10 \tab 20 \tab 25 \tab 18 \tab 10 \tab 13 \tab 12 \tab 14 \tab 16 \tab 19 \tab 21 \tab 13 \tab 11 \tab 19 \tab 12 \tab 15 \tab 15 \tab 14 \tab 8 \tab - \tab 2904 \cr
   1998 \tab 2832 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2000 \tab 2817 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 2812 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2812 \cr
   2006 \tab 2987 \tab 20 \tab 22 \tab 26 \tab 18 \tab 15 \tab 19 \tab 22 \tab 15 \tab 20 \tab 22 \tab 22 \tab 22 \tab 21 \tab 19 \tab 22 \tab 20 \tab 25 \tab 26 \tab 27 \tab 17 \tab 23 \tab 23 \tab 22 \tab 19 \tab 25 \tab 21 \tab 23 \tab 21 \tab 22 \tab 20 \tab 19 \tab 26 \tab 21 \tab 19 \tab 21 \tab 18 \tab 19 \tab 19 \tab 21 \tab 22 \tab 22 \tab 16 \tab 17 \tab 24 \tab 21 \tab 19 \tab 24 \tab 24 \tab 26 \tab - \tab - \tab - \tab - \tab - \tab 17 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 25 \tab - \tab - \tab - \tab 25 \tab 17 \tab 18 \tab 24 \tab 15 \tab 15 \tab 27 \tab 26 \tab 22 \tab 20 \tab 25 \tab 20 \tab 23 \tab 20 \tab 22 \tab 21 \tab 23 \tab 18 \tab 20 \tab 21 \tab 22 \tab - \tab 4510 \cr
   2008 \tab 2023 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2023 \cr
   2010 \tab 2044 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2044 \cr
   2012 \tab 1974 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 1974 \cr
   2014 \tab 2538 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2538 \cr
   2016 \tab 2867 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2867 \cr
   2018 \tab 1175 \tab 12 \tab 14 \tab 12 \tab 14 \tab 15 \tab 12 \tab 17 \tab 14 \tab 12 \tab 11 \tab 12 \tab 18 \tab 12 \tab 12 \tab 17 \tab 10 \tab 16 \tab 17 \tab 13 \tab 11 \tab 13 \tab 16 \tab 12 \tab 12 \tab 14 \tab 18 \tab 19 \tab 12 \tab 14 \tab 14 \tab 14 \tab 11 \tab 13 \tab 12 \tab 9 \tab 14 \tab 12 \tab 14 \tab 8 \tab 8 \tab 11 \tab 15 \tab 14 \tab 19 \tab 12 \tab 17 \tab 12 \tab 14 \tab 18 \tab 14 \tab 10 \tab 12 \tab 9 \tab 11 \tab 15 \tab 16 \tab 16 \tab 12 \tab 12 \tab 9 \tab 15 \tab 11 \tab 13 \tab 15 \tab 15 \tab 16 \tab 12 \tab 15 \tab 10 \tab 18 \tab 12 \tab 13 \tab 13 \tab 16 \tab 11 \tab 12 \tab 14 \tab 9 \tab 10 \tab 9 \tab 12 \tab 12 \tab 9 \tab 9 \tab 9 \tab 12 \tab 12 \tab 9 \tab 19 \tab 13 \tab - \tab 2348 \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 3544 \tab 3544 \cr
   2024 \tab 1670 \tab 18 \tab 14 \tab 15 \tab 15 \tab 16 \tab 16 \tab 17 \tab 12 \tab 20 \tab 23 \tab 19 \tab 17 \tab 14 \tab 21 \tab 18 \tab 15 \tab 21 \tab 15 \tab 13 \tab 26 \tab 10 \tab 28 \tab 19 \tab 15 \tab 28 \tab 21 \tab 10 \tab 17 \tab 19 \tab 22 \tab 17 \tab 19 \tab 16 \tab 16 \tab 19 \tab 17 \tab 13 \tab 13 \tab 15 \tab 23 \tab 21 \tab 19 \tab 18 \tab 24 \tab 26 \tab 21 \tab 28 \tab 17 \tab 16 \tab 20 \tab 20 \tab 21 \tab 25 \tab 19 \tab 14 \tab 21 \tab 15 \tab 21 \tab 18 \tab 10 \tab 21 \tab 21 \tab 20 \tab 15 \tab 14 \tab 23 \tab 14 \tab 16 \tab 20 \tab 14 \tab 22 \tab 13 \tab 17 \tab 18 \tab 18 \tab 27 \tab 17 \tab 15 \tab 21 \tab 21 \tab 15 \tab 18 \tab 15 \tab 15 \tab 19 \tab 24 \tab 16 \tab 14 \tab 15 \tab 25 \tab - \tab 3309 \cr
   Total \tab 66376 \tab 72 \tab 68 \tab 71 \tab 61 \tab 58 \tab 58 \tab 74 \tab 61 \tab 74 \tab 70 \tab 68 \tab 67 \tab 63 \tab 73 \tab 74 \tab 64 \tab 77 \tab 73 \tab 67 \tab 77 \tab 67 \tab 78 \tab 65 \tab 65 \tab 84 \tab 68 \tab 76 \tab 66 \tab 74 \tab 75 \tab 69 \tab 72 \tab 63 \tab 64 \tab 69 \tab 69 \tab 60 \tab 58 \tab 56 \tab 72 \tab 73 \tab 71 \tab 59 \tab 84 \tab 70 \tab 75 \tab 82 \tab 72 \tab 77 \tab 49 \tab 44 \tab 41 \tab 51 \tab 48 \tab 62 \tab 55 \tab 46 \tab 48 \tab 51 \tab 36 \tab 52 \tab 53 \tab 46 \tab 47 \tab 43 \tab 81 \tab 42 \tab 50 \tab 48 \tab 66 \tab 64 \tab 54 \tab 74 \tab 74 \tab 62 \tab 76 \tab 70 \tab 58 \tab 65 \tab 71 \tab 66 \tab 74 \tab 57 \tab 57 \tab 68 \tab 71 \tab 61 \tab 58 \tab 69 \tab 68 \tab 3544 \tab 75699 \cr
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2018 \tab B/C/- \tab partial \cr
   2024 \tab A/B/- \tab partial \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{charactr}},
\code{\link{imbalnce}},
\code{\link{probfix3}},
\code{\link{probhlp1}},
\code{\link{probhlp2}},
\code{\link{probhlp3}},
\code{\link{seriousp}},
\code{\link{stresses}},
\code{\link{wayraise}}

Other Mental Health: 
\code{\link{charactr}},
\code{\link{imbalnce}},
\code{\link{probfix3}},
\code{\link{probhlp1}},
\code{\link{probhlp2}},
\code{\link{probhlp3}},
\code{\link{seriousp}},
\code{\link{stresses}},
\code{\link{wayraise}}
}
\concept{Mental Health}
\keyword{variable}
