% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_210.R
\name{vigversn}
\alias{vigversn}
\title{Vignette version number}
\source{
General Social Survey https://gss.norc.org
}
\description{
vigversn
}
\details{
Question Next I'm going to describe a person -- lets call him/her (REFER TO NAME IN VIGNETTE). After I read a description of him/her I'll ask you some questions about how you think and feel about him/her. There are no right or wrong answers. I'm only interested in what you think of NAME.
}
\section{Values}{

\itemize{
\item \code{1} 1: alcohol dependence, john, white, eighth grade
\item \code{2} 2: alcohol dependence, john, african american, eighth grade
\item \code{3} 3: alcohol dependence, john, hispanic, eighth grade
\item \code{4} 4: alcohol dependence, john, white, high school
\item \code{5} 5: alcohol dependence, john, african american, high school
\item \code{6} 6: alcohol dependence, john, hispanic, high school
\item \code{7} 7: alcohol dependence, john, white, college
\item \code{8} 8: alcohol dependence, john, african american, college
\item \code{9} 9: alcohol dependence, john, hispanic, college
\item \code{10} 10: alcohol dependence, mary, white, eighth grade
\item \code{11} 11: alcohol dependence, mary, african american, eighth grade
\item \code{12} 12: alcohol dependence, mary, hispanic, eighth grade
\item \code{13} 13: alcohol dependence, mary, white, high school
\item \code{14} 14: alcohol dependence, mary, african american, high school
\item \code{15} 15: alcohol dependence, mary, hispanic, high school
\item \code{16} 16: alcohol dependence, mary, white, college
\item \code{17} 17: alcohol dependence, mary, african american, college
\item \code{18} 18: alcohol dependence, mary, hispanic, college
\item \code{19} 19: major depression, john, white, eighth grade
\item \code{20} 20: major depression, john, african american, eighth grade
\item \code{21} 21: major depression, john, hispanic, eighth grade
\item \code{22} 22: major depression, john, white, high school
\item \code{23} 23: major depression, john, african american, high school
\item \code{24} 24: major depression, john, hispanic, high school
\item \code{25} 25: major depression, john, white, college
\item \code{26} 26: major depression, john, african american, college
\item \code{27} 27: major depression, john, hispanic, college
\item \code{28} 28: major depression, mary, white, eighth grade
\item \code{29} 29: major depression, mary, african american, eighth grade
\item \code{30} 30: major depression, mary, hispanic, eighth grade
\item \code{31} 31: major depression, mary, white, high school
\item \code{32} 32: major depression, mary, african american, high school
\item \code{33} 33: major depression, mary, hispanic, high school
\item \code{34} 34: major depression, mary, white, college
\item \code{35} 35: major depression, mary, african american, college
\item \code{36} 36: major depression, mary, hispanic, college
\item \code{37} 37: schizophrenia, john, white, eighth grade
\item \code{38} 38: schizophrenia, john, african american, eighth grade
\item \code{39} 39: schizophrenia, john, hispanic, eighth grade
\item \code{40} 40: schizophrenia, john, white, high school
\item \code{41} 41: schizophrenia, john, african american, high school
\item \code{42} 42: schizophrenia, john, hispanic, high school
\item \code{43} 43: schizophrenia, john, white, college
\item \code{44} 44: schizophrenia, john, african american, college
\item \code{45} 45: schizophrenia, john, hispanic, college
\item \code{46} 46: schizophrenia, mary, white, eighth grade
\item \code{47} 47: schizophrenia, mary, african american, eighth grade
\item \code{48} 48: schizophrenia, mary, hispanic, eighth grade
\item \code{49} 49: schizophrenia, mary, white, high school
\item \code{50} 50: schizophrenia, mary, african american, high school
\item \code{51} 51: schizophrenia, mary, hispanic, high school
\item \code{52} 52: schizophrenia, mary, white, college
\item \code{53} 53: schizophrenia, mary, african american, college
\item \code{54} 54: schizophrenia, mary, hispanic, college
\item \code{55} 55: drug problem, john, white, eighth grade
\item \code{56} 56: drug problem, john, african american, eighth grade
\item \code{57} 57: drug problem, john, hispanic, eighth grade
\item \code{58} 58: drug problem, john, white, high school
\item \code{59} 59: drug problem, john, african american, high school
\item \code{60} 60: drug problem, john, hispanic, high school
\item \code{61} 61: drug problem, john, white, college
\item \code{62} 62: drug problem, john, african american, college
\item \code{63} 63: drug problem, john, hispanic, college
\item \code{64} 64: drug problem, mary, white, eighth grade
\item \code{65} 65: drug problem, mary, african american, eighth grade
\item \code{66} 66: drug problem, mary, hispanic, eighth grade
\item \code{67} 67: drug problem, mary, white, high school
\item \code{68} 68: drug problem, mary, african american, high school
\item \code{69} 69: drug problem, mary, hispanic, high school
\item \code{70} 70: drug problem, mary, white, college
\item \code{71} 71: drug problem, mary, african american, college
\item \code{72} 72: drug problem, mary, hispanic, college
\item \code{73} 73: no problem, john, white, eighth grade
\item \code{74} 74: no problem, john, african american, eighth grade
\item \code{75} 75: no problem, john, hispanic, eighth grade
\item \code{76} 76: no problem, john, white, high school
\item \code{77} 77: no problem, john, african american, high school
\item \code{78} 78: no problem, john, hispanic, high school
\item \code{79} 79: no problem, john, white, college
\item \code{80} 80: no problem, john, african american, college
\item \code{81} 81: no problem, john, hispanic, college
\item \code{82} 82: no problem, mary, white, eighth grade
\item \code{83} 83: no problem, mary, african american, eighth grade
\item \code{84} 84: no problem, mary, hispanic, eighth grade
\item \code{85} 85: no problem, mary, white, high school
\item \code{86} 86: no problem, mary, african american, high school
\item \code{87} 87: no problem, mary, hispanic, high school
\item \code{88} 88: no problem, mary, white, college
\item \code{89} 89: no problem, mary, african american, college
\item \code{90} 90: no problem, mary, hispanic, college
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2018 \tab B/C/- \tab partial \cr
   2024 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1986/vshow}{GSS Data Explorer page for \code{vigversn}}.
}

\seealso{
Other Mental Health: 
\code{\link{anydoc}},
\code{\link{breakdwn}},
\code{\link{brkdas1}},
\code{\link{brkdas2}},
\code{\link{brkddo1}},
\code{\link{brkddo2}},
\code{\link{brkddo3}},
\code{\link{brkddur}},
\code{\link{brkdhlp1}},
\code{\link{brkdhlp2}},
\code{\link{brkdhlp3}},
\code{\link{brkdimp1}},
\code{\link{brkdimp2}},
\code{\link{brkdimp3}},
\code{\link{brkdoth}},
\code{\link{brkdsym1}},
\code{\link{brkdsym2}},
\code{\link{brkdsym3}},
\code{\link{brkdtime}},
\code{\link{brkdwhy1}},
\code{\link{brkdwhy2}},
\code{\link{brkdwhy3}},
\code{\link{bullied}},
\code{\link{bullpeer}},
\code{\link{charactr}},
\code{\link{chldadlt_asn}},
\code{\link{chldadlt_blk}},
\code{\link{chldadlt_hisp}},
\code{\link{chldadlt_oth}},
\code{\link{chldadlt_wht}},
\code{\link{chldchld_asn}},
\code{\link{chldchld_blk}},
\code{\link{chldchld_hisp}},
\code{\link{chldchld_oth}},
\code{\link{chldchld_wht}},
\code{\link{chldfam_asn}},
\code{\link{chldfam_blk}},
\code{\link{chldfam_hisp}},
\code{\link{chldfam_oth}},
\code{\link{chldfam_wht}},
\code{\link{chldshop_asn}},
\code{\link{chldshop_blk}},
\code{\link{chldshop_hisp}},
\code{\link{chldshop_oth}},
\code{\link{chldshop_wht}},
\code{\link{chldvig}},
\code{\link{cutself}},
\code{\link{dangroth}},
\code{\link{dangrslf}},
\code{\link{decmoney}},
\code{\link{dectreat}},
\code{\link{diagnosd}},
\code{\link{doc1}},
\code{\link{doc10}},
\code{\link{doc11}},
\code{\link{doc12}},
\code{\link{doc13}},
\code{\link{doc14}},
\code{\link{doc15}},
\code{\link{doc16}},
\code{\link{doc17}},
\code{\link{doc18}},
\code{\link{doc19}},
\code{\link{doc2}},
\code{\link{doc20}},
\code{\link{doc3}},
\code{\link{doc4}},
\code{\link{doc5}},
\code{\link{doc6}},
\code{\link{doc7}},
\code{\link{doc8}},
\code{\link{doc9}},
\code{\link{doclist}},
\code{\link{docsui}},
\code{\link{dofirst}},
\code{\link{evbrkdwn}},
\code{\link{evmhp}},
\code{\link{fammhneg}},
\code{\link{fighpeer}},
\code{\link{foodalc}},
\code{\link{genetics}},
\code{\link{godswill}},
\code{\link{govmentl}},
\code{\link{hlthinsr}},
\code{\link{hmo1}},
\code{\link{hmo2}},
\code{\link{hmo3}},
\code{\link{hmo4}},
\code{\link{hmo5}},
\code{\link{hmo6}},
\code{\link{hmo7}},
\code{\link{hospice}},
\code{\link{hurtoth}},
\code{\link{hurtself}},
\code{\link{imbalnce}},
\code{\link{imprvown}},
\code{\link{imprvown1}},
\code{\link{imprvtrt}},
\code{\link{instrctn}},
\code{\link{knwmhosp}},
\code{\link{knwnproz}},
\code{\link{knwpatnt}},
\code{\link{medcare1}},
\code{\link{medcare2}},
\code{\link{meddoc}},
\code{\link{mentldoc}},
\code{\link{mentlhos}},
\code{\link{mentlill}},
\code{\link{mentloth}},
\code{\link{mnildo1}},
\code{\link{mnildo2}},
\code{\link{mnildo3}},
\code{\link{mnildur}},
\code{\link{mnilhlp1}},
\code{\link{mnilhlp2}},
\code{\link{mnilhlp3}},
\code{\link{mniltime}},
\code{\link{mnilwhy1}},
\code{\link{mnilwhy2}},
\code{\link{mnilwhy3}},
\code{\link{mntlas1}},
\code{\link{mntlas2}},
\code{\link{mntlcare}},
\code{\link{mntldeny}},
\code{\link{mntldif}},
\code{\link{mntlimp1}},
\code{\link{mntlimp2}},
\code{\link{mntlimp3}},
\code{\link{mntloth}},
\code{\link{mntlsym1}},
\code{\link{mntlsym2}},
\code{\link{mntlsym3}},
\code{\link{morextrm}},
\code{\link{mustdoc}},
\code{\link{mustdoc1}},
\code{\link{musthosp}},
\code{\link{musthosp1}},
\code{\link{mustmed}},
\code{\link{mustmed1}},
\code{\link{negself}},
\code{\link{notreat}},
\code{\link{orhealer}},
\code{\link{ormeddoc}},
\code{\link{ormntldc}},
\code{\link{ormntlhs}},
\code{\link{ormntlot}},
\code{\link{orotcmed}},
\code{\link{orrxmed}},
\code{\link{orslfhlp}},
\code{\link{ortlkclr}},
\code{\link{ortlkfm}},
\code{\link{otcmed}},
\code{\link{othmhneg}},
\code{\link{painmed}},
\code{\link{physill}},
\code{\link{physill1}},
\code{\link{probfix1}},
\code{\link{probfix2}},
\code{\link{probfix3}},
\code{\link{probhlp1}},
\code{\link{probhlp2}},
\code{\link{probhlp3}},
\code{\link{proz1}},
\code{\link{proz2}},
\code{\link{proz3}},
\code{\link{proz4}},
\code{\link{proz5}},
\code{\link{proz6}},
\code{\link{proz7}},
\code{\link{proz8}},
\code{\link{proz9}},
\code{\link{prozac}},
\code{\link{prozfor1}},
\code{\link{prozfor2}},
\code{\link{prozfor3}},
\code{\link{prozkid1}},
\code{\link{prozkid2}},
\code{\link{prozkid3}},
\code{\link{prozslf1}},
\code{\link{prozslf2}},
\code{\link{prozslf3}},
\code{\link{prozslf4}},
\code{\link{psycdrct}},
\code{\link{psycfdoc}},
\code{\link{psyckid1}},
\code{\link{psyckid2}},
\code{\link{psyckid3}},
\code{\link{psycmed1}},
\code{\link{psycmed2}},
\code{\link{psycmed3}},
\code{\link{psycmed4}},
\code{\link{psycmed5}},
\code{\link{psycmed6}},
\code{\link{psycmed7}},
\code{\link{psycpsyc}},
\code{\link{relmhsp1}},
\code{\link{relmhsp2}},
\code{\link{relmhsp3}},
\code{\link{relmhsp4}},
\code{\link{relmhsp5}},
\code{\link{relmhsp6}},
\code{\link{rightref}},
\code{\link{risklife}},
\code{\link{rumopeer}},
\code{\link{rxmed}},
\code{\link{schshoot}},
\code{\link{seenmntl}},
\code{\link{selfhelp}},
\code{\link{seriousp}},
\code{\link{smedia}},
\code{\link{socsecfx}},
\code{\link{socsecnu}},
\code{\link{socsecrt}},
\code{\link{sphealer}},
\code{\link{spmentl}},
\code{\link{stresses}},
\code{\link{swithlth}},
\code{\link{takelife}},
\code{\link{termecon}},
\code{\link{termemot}},
\code{\link{terminal}},
\code{\link{termpain}},
\code{\link{termpay}},
\code{\link{termrel}},
\code{\link{termvig}},
\code{\link{tlkclrgy}},
\code{\link{tlkfam}},
\code{\link{trustcrt}},
\code{\link{trustdoc}},
\code{\link{trustfam}},
\code{\link{upsdowns}},
\code{\link{usedmntl}},
\code{\link{usedproz}},
\code{\link{usepsyc1}},
\code{\link{usepsyc2}},
\code{\link{usepsyc3}},
\code{\link{usepsyc4}},
\code{\link{vigfrnd}},
\code{\link{viggrp}},
\code{\link{viglabel}},
\code{\link{vigmar}},
\code{\link{vignei}},
\code{\link{vigsoc}},
\code{\link{vigwork}},
\code{\link{wayraise}}
}
\concept{Mental Health}
\keyword{variable}
