% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_128.R
\name{violblks}
\alias{violblks}
\title{Violent - not violent}
\source{
General Social Survey https://gss.norc.org
}
\description{
violblks
}
\details{
Question Now I have some questions about different groups in our society. I'm going to show you a seven-point scale on which the characteristics of people in a group can be rated. In the first statement a score of 1 means that you think almost all of the people in that group are "rich."  A score of 7 means that you think almost everyone in the group are "poor."  A score of 4 means you think that the group is not towards one end or another, and of course you may choose any number In between that comes closest to where you think people in the group stand.
C. The next set asks if people in each group tend to be violence prone or if they tend not to be prone to violence?
3. Blacks?
}
\section{Values}{

\itemize{
\item \code{1} violence-prone
\item \code{7} not violence-prone
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1192/vshow}{GSS Data Explorer page for \code{violblks}}.
}

\seealso{
Other Intergroup Relations: 
\code{\link{intlasns}},
\code{\link{intlblks}},
\code{\link{intljews}},
\code{\link{intlwhts}},
\code{\link{violasns}},
\code{\link{violhsps}},
\code{\link{violjews}},
\code{\link{violso}},
\code{\link{violwhts}}

Other Blacks: 
\code{\link{intlblks}}
}
\concept{Blacks}
\concept{Intergroup Relations}
\keyword{variable}
