% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_331.R
\name{violpeop}
\alias{violpeop}
\title{Would child do something violent toward other}
\source{
General Social Survey https://gss.norc.org
}
\description{
violpeop
}
\details{
Question A. How likely in your opinion is it (John/Mary) would do something violent toward other people -- is it very likely, somewhat likely, not very likely, or not at all likely?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} not very likely
\item \code{4} not likely at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   A/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2955/vshow}{GSS Data Explorer page for \code{violpeop}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not likely at all \tab don't know \tab no answer \tab Total \cr
   2002 \tab 1372 \tab 64 \tab 266 \tab 613 \tab 389 \tab 48 \tab 13 \tab 2765 \cr
   Total \tab 1372 \tab 64 \tab 266 \tab 613 \tab 389 \tab 48 \tab 13 \tab 2765 \cr
}
}

\seealso{
Other Mental Health II: 
\code{\link{comknows}},
\code{\link{failure}},
\code{\link{illphys}},
\code{\link{medkdneg}},
\code{\link{outsider}},
\code{\link{ovrmedkd}},
\code{\link{putsoff}},
\code{\link{sufadult}},
\code{\link{violself}}
}
\concept{Mental Health II}
\keyword{variable}
