% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_426.R
\name{visa}
\alias{visa}
\title{Do you hold a special visa}
\source{
General Social Survey https://gss.norc.org
}
\description{
visa
}
\details{
Question Do you hold a special visa?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3900/vshow}{GSS Data Explorer page for \code{visa}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1838 \tab 41 \tab 135 \tab 3 \tab 6 \tab 2023 \cr
   Total \tab 1838 \tab 41 \tab 135 \tab 3 \tab 6 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{agecmeus}},
\code{\link{bornsp}},
\code{\link{emphlth}},
\code{\link{emphplan}},
\code{\link{natdeg}},
\code{\link{othplan}},
\code{\link{typevisa}}
}
\concept{Global Economics}
\keyword{variable}
