% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_517.R
\name{vistholy}
\alias{vistholy}
\title{How often r visits holy place for religious reasons}
\source{
General Social Survey https://gss.norc.org
}
\description{
vistholy
}
\details{
Question How often do you visit a holy place for religious reasons such as going to a shrine, temple, church, or mosque?
Please do not count attending regular religious services at your usual place of worship, if you have one.
Would you say…
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} less than once a year
\item \code{3} about once or twice a year
\item \code{4} several times a year
\item \code{5} about once a month or more
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab B/C \tab partial \cr
   2018 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4710/vshow}{GSS Data Explorer page for \code{vistholy}}.
}

\seealso{
Other ISSP Religion: 
\code{\link{comfort}},
\code{\link{makefrnd}},
\code{\link{paxhappy}},
\code{\link{relobjct}},
\code{\link{relsprt}},
\code{\link{rightpeo}}
}
\concept{ISSP Religion}
\keyword{variable}
