% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_12.R
\name{voedncol}
\alias{voedncol}
\title{Non-college postsecondary education (voednme2)}
\source{
General Social Survey https://gss.norc.org
}
\description{
voedncol
}
\details{
Question In what year did you receive the (COLLEGE NAME) degree?
G. Did you complete any courses for credit at an institution such as a business college, technical or vocational school?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2012, 2014, 2016, 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/75/vshow}{GSS Data Explorer page for \code{voedncol}}.

Counts by year:\tabular{lllll}{
   year \tab iap \tab yes \tab no \tab Total \cr
   2012 \tab 1227 \tab 244 \tab 503 \tab 1974 \cr
   2014 \tab 1573 \tab 288 \tab 677 \tab 2538 \cr
   2016 \tab 1779 \tab 350 \tab 738 \tab 2867 \cr
   2018 \tab 2348 \tab - \tab - \tab 2348 \cr
   Total \tab 6927 \tab 882 \tab 1918 \tab 9727 \cr
}
}

\seealso{
Other Core: 
\code{\link{barate}},
\code{\link{eftotlt}},
\code{\link{gradtounder}},
\code{\link{sector}},
\code{\link{speftotlt}},
\code{\link{spsector}},
\code{\link{voedcol}},
\code{\link{voednme1}},
\code{\link{voednme2}}
}
\concept{Core}
\keyword{variable}
