% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_12.R
\name{voednme1}
\alias{voednme1}
\title{Postsecondary institution attended for credit}
\source{
General Social Survey https://gss.norc.org
}
\description{
voednme1
}
\details{
Question In what year did you receive the (COLLEGE NAME) degree?
F. IF YES: What institution was that?
}
\section{Values}{

\itemize{
\item \code{1} jail/prison
\item \code{2} high school
\item \code{3} company/business
\item \code{4} vocational school/trades/technical (not high school, not college)
\item \code{5} secretarial school
\item \code{6} adult education/school
\item \code{7} medical/nursing/health/dental/hospital
\item \code{8} union
\item \code{9} armed forces
\item \code{10} college/university
\item \code{11} art/design
\item \code{12} business college (not business school at a university)
\item \code{13} home study/correspondence
\item \code{14} job corp
\item \code{15} beauty school/cosmetology
\item \code{16} insurance, real estate, finance, executive management
\item \code{17} police/fire academy; training
\item \code{18} development center
\item \code{19} career center
\item \code{20} foreign college/university
\item \code{30} other
\item \code{40} unable to code response, unclear, not identifiable
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2016 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/74/vshow}{GSS Data Explorer page for \code{voednme1}}.
}

\seealso{
Other Core: 
\code{\link{barate}},
\code{\link{eftotlt}},
\code{\link{gradtounder}},
\code{\link{sector}},
\code{\link{speftotlt}},
\code{\link{spsector}},
\code{\link{voedcol}},
\code{\link{voedncol}},
\code{\link{voednme2}}
}
\concept{Core}
\keyword{variable}
