% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_433.R
\name{volacty2}
\alias{volacty2}
\title{Done other types of volunteering for child's school or youth org}
\source{
General Social Survey https://gss.norc.org
}
\description{
volacty2
}
\details{
Question Sometimes people don’t think of activities they do infrequently or activities they do for children’s schools or youth organizations as volunteer activities. Since (CURRENT MONTH) 1st of last year, have you done any of these types of volunteer activities?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3971/vshow}{GSS Data Explorer page for \code{volacty2}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2012 \tab 1192 \tab 96 \tab 681 \tab 5 \tab 1974 \cr
   Total \tab 1192 \tab 96 \tab 681 \tab 5 \tab 1974 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{polefy11}},
\code{\link{polefy13}},
\code{\link{polefy15}},
\code{\link{polefy16}},
\code{\link{polefy17}},
\code{\link{polefy3}},
\code{\link{prespop}},
\code{\link{volactyr}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
