% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_433.R
\name{volactyr}
\alias{volactyr}
\title{Since last yr any volunteering}
\source{
General Social Survey https://gss.norc.org
}
\description{
volactyr
}
\details{
Question Since (CURRENT MONTH) 1st of last year, have you done any volunteer activities through or for an organization?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3970/vshow}{GSS Data Explorer page for \code{volactyr}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2012 \tab 675 \tab 517 \tab 777 \tab 5 \tab 1974 \cr
   Total \tab 675 \tab 517 \tab 777 \tab 5 \tab 1974 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{polefy11}},
\code{\link{polefy13}},
\code{\link{polefy15}},
\code{\link{polefy16}},
\code{\link{polefy17}},
\code{\link{polefy3}},
\code{\link{prespop}},
\code{\link{volacty2}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
