% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_325.R
\name{volmonth}
\alias{volmonth}
\title{Volunteer in last month}
\source{
General Social Survey https://gss.norc.org
}
\description{
volmonth
}
\details{
Question In the past month, did you do any volunteer activity through organizations, that is donate your time and energy not for pay?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
   A/C/- \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2894/vshow}{GSS Data Explorer page for \code{volmonth}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2012 \tab 666 \tab 449 \tab 847 \tab 2 \tab 10 \tab 1974 \cr
   2014 \tab 1264 \tab 446 \tab 821 \tab 3 \tab 4 \tab 2538 \cr
   Total \tab 1930 \tab 895 \tab 1668 \tab 5 \tab 14 \tab 4512 \cr
}
}

\seealso{
Other Altruism: 
\code{\link{valgiven}}
}
\concept{Altruism}
\keyword{variable}
