% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_515.R
\name{volwkoth}
\alias{volwkoth}
\title{Volunteered other kinds past 12 month}
\source{
General Social Survey https://gss.norc.org
}
\description{
volwkoth
}
\details{
Question Have you done any voluntary activity in the past 12 months in any of the following areas? Voluntary activity is unpaid work, not just belonging to an organization or group. It should be of service or benefit to other people or the community and not only to one's family or personal friends. During the last 12 months did you do volunteer work in any of the following areas:
D. Any other kind of voluntary activities
}
\section{Values}{

\itemize{
\item \verb{[1]} no
\item \verb{[2]} yes 1-2 times
\item \verb{[3]} yes 3-5 times
\item \verb{[4]} yes 6/more times
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4694/vshow}{GSS Data Explorer page for \code{volwkoth}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab no \tab yes 1-2 times \tab yes 3-5 times \tab yes 6/more times \tab no answer \tab Total \cr
   1998 \tab 1548 \tab 755 \tab 222 \tab 109 \tab 138 \tab 60 \tab 2832 \cr
   Total \tab 1548 \tab 755 \tab 222 \tab 109 \tab 138 \tab 60 \tab 2832 \cr
}
}

\seealso{
Other ISSP Religion: 
\code{\link{cardo}},
\code{\link{carright}},
\code{\link{geomobil}},
\code{\link{relgrpeq}},
\code{\link{reltruth}},
\code{\link{rspctrel}},
\code{\link{volwkchr}},
\code{\link{volwkpol}},
\code{\link{volwkrel}}
}
\concept{ISSP Religion}
\keyword{variable}
