% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_22.R
\name{vote00}
\alias{vote00}
\title{Did r vote in 2000 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
vote00
}
\details{
Question I. In 2000, you remember that Gore ran for President on the Democratic ticket against Bush for the Republicans . Do you remember for sure whether or not you voted in that election?
}
\section{Values}{

\itemize{
\item \code{1} voted
\item \code{2} did not vote
\item \code{3} ineligible
\item \code{4} refused to answer
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2004, 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/166/vshow}{GSS Data Explorer page for \code{vote00}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab voted \tab did not vote \tab ineligible \tab refused to answer \tab don't know \tab no answer \tab Total \cr
   2002 \tab - \tab 1780 \tab 822 \tab 138 \tab 11 \tab 14 \tab - \tab 2765 \cr
   2004 \tab - \tab 1760 \tab 814 \tab 211 \tab 4 \tab 23 \tab - \tab 2812 \cr
   2006 \tab 1518 \tab 1826 \tab 715 \tab 389 \tab - \tab 52 \tab 10 \tab 4510 \cr
   Total \tab 1518 \tab 5366 \tab 2351 \tab 738 \tab 15 \tab 89 \tab 10 \tab 10087 \cr
}
}

\seealso{
Other Core: 
\code{\link{if00who}},
\code{\link{if04who}},
\code{\link{if96who}},
\code{\link{pres00}},
\code{\link{pres04}},
\code{\link{pres08}},
\code{\link{pres96}},
\code{\link{vote04}},
\code{\link{vote08}}
}
\concept{Core}
\keyword{variable}
