% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_22.R
\name{vote04}
\alias{vote04}
\title{Did r vote in 2004 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
vote04
}
\details{
Question J. In 2004, you remember that Kerry ran for President on the Democratic ticket against Bush for the Republicans. Do you remember for sure whether or not you voted in that election?
}
\section{Values}{

\itemize{
\item \code{1} voted
\item \code{2} did not vote
\item \code{3} ineligible
\item \code{4} refused to answer
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 2008, 2010 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/169/vshow}{GSS Data Explorer page for \code{vote04}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab voted \tab did not vote \tab ineligible \tab don't know \tab no answer \tab Total \cr
   2006 \tab - \tab 3037 \tab 1089 \tab 335 \tab 34 \tab 15 \tab 4510 \cr
   2008 \tab - \tab 1248 \tab 539 \tab 175 \tab 54 \tab 7 \tab 2023 \cr
   2010 \tab - \tab 1220 \tab 534 \tab 231 \tab 46 \tab 13 \tab 2044 \cr
   Total \tab 0 \tab 5505 \tab 2162 \tab 741 \tab 134 \tab 35 \tab 8577 \cr
}
}

\seealso{
Other Core: 
\code{\link{if00who}},
\code{\link{if04who}},
\code{\link{if96who}},
\code{\link{pres00}},
\code{\link{pres04}},
\code{\link{pres08}},
\code{\link{pres96}},
\code{\link{vote00}},
\code{\link{vote08}}
}
\concept{Core}
\keyword{variable}
