% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_23.R
\name{vote16}
\alias{vote16}
\title{Remember if voted in 2016 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
vote16
}
\details{
Question In 2016, you remember that Clinton ran for President on the Democratic ticket against Trump for the Republicans.
Do you remember for sure whether or not you voted in that election?
}
\section{Values}{

\itemize{
\item \code{1} voted
\item \code{2} did not vote
\item \code{3} ineligible
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7089/vshow}{GSS Data Explorer page for \code{vote16}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab voted \tab did not vote \tab ineligible \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2018 \tab - \tab 1523 \tab 658 \tab 132 \tab 21 \tab 14 \tab - \tab 2348 \cr
   2021 \tab - \tab 2886 \tab 596 \tab 221 \tab 80 \tab 244 \tab 5 \tab 4032 \cr
   2022 \tab - \tab 2219 \tab 706 \tab 356 \tab 146 \tab 112 \tab 5 \tab 3544 \cr
   Total \tab 0 \tab 6628 \tab 1960 \tab 709 \tab 247 \tab 370 \tab 10 \tab 9924 \cr
}
}

\keyword{variable}
