% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_20.R
\name{vote72}
\alias{vote72}
\title{Did r vote in 1972 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
vote72
}
\details{
Question B. In 1972, you remember that McGovern ran for President on the Democratic ticket against Nixon for the Republicans. Do you remember for sure whether or not you voted in that election?
}
\section{Values}{

\itemize{
\item \code{1} voted
\item \code{2} did not vote
\item \code{3} not eligible
\item \code{4} refused
\item \code{6} wldnt vt-relig
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1973, 1974, 1975, 1976, 1977 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/145/vshow}{GSS Data Explorer page for \code{vote72}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab voted \tab did not vote \tab not eligible \tab refused \tab wldnt vt-relig \tab don't know \tab no answer \tab Total \cr
   1973 \tab - \tab 1045 \tab 422 \tab 29 \tab 3 \tab 2 \tab 1 \tab 2 \tab 1504 \cr
   1974 \tab - \tab 1013 \tab 416 \tab 35 \tab 6 \tab - \tab 13 \tab 1 \tab 1484 \cr
   1975 \tab - \tab 960 \tab 415 \tab 82 \tab 3 \tab - \tab 26 \tab 4 \tab 1490 \cr
   1976 \tab - \tab 950 \tab 402 \tab 122 \tab 3 \tab - \tab 22 \tab - \tab 1499 \cr
   1977 \tab - \tab 915 \tab 390 \tab 188 \tab 3 \tab - \tab 29 \tab 5 \tab 1530 \cr
   Total \tab 0 \tab 4883 \tab 2045 \tab 456 \tab 18 \tab 2 \tab 91 \tab 12 \tab 7507 \cr
}
}

\seealso{
Other Core: 
\code{\link{if68who}},
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{if80who}},
\code{\link{pres72}},
\code{\link{pres76}},
\code{\link{pres80}},
\code{\link{vote76}},
\code{\link{vote80}}
}
\concept{Core}
\keyword{variable}
