% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_20.R
\name{vote76}
\alias{vote76}
\title{Did r vote in 1976 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
vote76
}
\details{
Question C. In 1976, you remember that Carter ran for President on the Democratic ticket against Ford for the Republicans. Do you remember for sure whether or not you voted in that election?
}
\section{Values}{

\itemize{
\item \verb{[1]} voted
\item \verb{[2]} did not vote
\item \verb{[3]} not eligible
\item \verb{[4]} refused
\item \verb{[6]} wldnt vt-relig
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1977, 1978, 1980, 1982 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/148/vshow}{GSS Data Explorer page for \code{vote76}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab voted \tab did not vote \tab not eligible \tab refused \tab don't know \tab no answer \tab Total \cr
   1977 \tab - \tab 983 \tab 479 \tab 56 \tab 3 \tab 4 \tab 5 \tab 1530 \cr
   1978 \tab - \tab 973 \tab 456 \tab 82 \tab 3 \tab 16 \tab 2 \tab 1532 \cr
   1980 \tab - \tab 896 \tab 440 \tab 107 \tab - \tab 22 \tab 3 \tab 1468 \cr
   1982 \tab - \tab 1100 \tab 477 \tab 221 \tab 5 \tab 54 \tab 3 \tab 1860 \cr
   Total \tab 0 \tab 3952 \tab 1852 \tab 466 \tab 11 \tab 96 \tab 13 \tab 6390 \cr
}
}

\seealso{
Other Core: 
\code{\link{if68who}},
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{if80who}},
\code{\link{pres72}},
\code{\link{pres76}},
\code{\link{pres80}},
\code{\link{vote72}},
\code{\link{vote80}}
}
\concept{Core}
\keyword{variable}
