% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_20.R
\name{vote80}
\alias{vote80}
\title{Did r vote in 1980 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
vote80
}
\details{
Question D. In 1980, you remember that Carter ran for President on the Democratic ticket against Reagan for the Republicans, and  anderson as an Independent. Do you remember for sure whether or not you voted in that election?
}
\section{Values}{

\itemize{
\item \verb{[1]} voted
\item \verb{[2]} did not vote
\item \verb{[3]} not eligible
\item \verb{[4]} refused
\item \verb{[6]} wldnt vt-relig
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1982, 1983, 1984, 1985, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/151/vshow}{GSS Data Explorer page for \code{vote80}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab voted \tab did not vote \tab not eligible \tab refused \tab don't know \tab no answer \tab Total \cr
   1982 \tab - \tab 1184 \tab 581 \tab 66 \tab 7 \tab 20 \tab 2 \tab 1860 \cr
   1983 \tab - \tab 1066 \tab 437 \tab 73 \tab 6 \tab 15 \tab 2 \tab 1599 \cr
   1984 \tab - \tab 945 \tab 391 \tab 106 \tab 3 \tab 19 \tab 9 \tab 1473 \cr
   1985 \tab - \tab 1019 \tab 359 \tab 130 \tab - \tab 24 \tab 2 \tab 1534 \cr
   1987 \tab - \tab 1139 \tab 373 \tab 240 \tab 5 \tab 53 \tab 9 \tab 1819 \cr
   Total \tab 0 \tab 5353 \tab 2141 \tab 615 \tab 21 \tab 131 \tab 24 \tab 8285 \cr
}
}

\seealso{
Other Core: 
\code{\link{if68who}},
\code{\link{if72who}},
\code{\link{if76who}},
\code{\link{if80who}},
\code{\link{pres72}},
\code{\link{pres76}},
\code{\link{pres80}},
\code{\link{vote72}},
\code{\link{vote76}}
}
\concept{Core}
\keyword{variable}
