% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_21.R
\name{vote84}
\alias{vote84}
\title{Did r vote in 1984 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
vote84
}
\details{
Question E. In 1984, you remember that Mondale ran for President on the Democratic ticket against Reagan for the Republicans. Do you remember for sure whether or not you voted in that election?
}
\section{Values}{

\itemize{
\item \code{1} voted
\item \code{2} did not vote
\item \code{3} not eligible
\item \code{4} refused
\item \code{6} wldnt vt-relig
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1988, 1989 \cr
   Full \tab 1985, 1986, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/154/vshow}{GSS Data Explorer page for \code{vote84}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab voted \tab did not vote \tab not eligible \tab refused \tab don't know \tab no answer \tab Total \cr
   1985 \tab - \tab 1057 \tab 429 \tab 36 \tab 1 \tab 9 \tab 2 \tab 1534 \cr
   1986 \tab - \tab 983 \tab 411 \tab 49 \tab 4 \tab 23 \tab - \tab 1470 \cr
   1987 \tab - \tab 1169 \tab 498 \tab 99 \tab 3 \tab 42 \tab 8 \tab 1819 \cr
   1988 \tab - \tab 905 \tab 416 \tab 117 \tab 5 \tab 36 \tab 2 \tab 1481 \cr
   1989 \tab - \tab 979 \tab 366 \tab 141 \tab 15 \tab 33 \tab 3 \tab 1537 \cr
   Total \tab 0 \tab 5093 \tab 2120 \tab 442 \tab 28 \tab 143 \tab 15 \tab 7841 \cr
}
}

\seealso{
Other Core: 
\code{\link{if84who}},
\code{\link{if88who}},
\code{\link{if92who}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{pres92}},
\code{\link{vote88}},
\code{\link{vote92}},
\code{\link{vote96}}
}
\concept{Core}
\keyword{variable}
