% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_21.R
\name{vote92}
\alias{vote92}
\title{Did r vote in 1992 election}
\source{
General Social Survey https://gss.norc.org
}
\description{
vote92
}
\details{
Question G. In 1992, you remember that Clinton ran for President on the Democratic ticket against Bush for the Republicans and Perot as an Independent.
Do you remember for sure whether or not you voted in that election?
}
\section{Values}{

\itemize{
\item \verb{[1]} voted
\item \verb{[2]} did not vote
\item \verb{[3]} not eligible
\item \verb{[4]} refused
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1993, 1994, 1996, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/160/vshow}{GSS Data Explorer page for \code{vote92}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab voted \tab did not vote \tab not eligible \tab refused \tab don't know \tab no answer \tab Total \cr
   1993 \tab - \tab 1109 \tab 446 \tab 36 \tab 6 \tab 5 \tab 4 \tab 1606 \cr
   1994 \tab - \tab 2053 \tab 818 \tab 81 \tab 14 \tab 23 \tab 3 \tab 2992 \cr
   1996 \tab - \tab 1907 \tab 763 \tab 184 \tab 10 \tab 38 \tab 2 \tab 2904 \cr
   1998 \tab - \tab 1767 \tab 655 \tab 284 \tab 13 \tab 82 \tab 31 \tab 2832 \cr
   Total \tab 0 \tab 6836 \tab 2682 \tab 585 \tab 43 \tab 148 \tab 40 \tab 10334 \cr
}
}

\seealso{
Other Core: 
\code{\link{if84who}},
\code{\link{if88who}},
\code{\link{if92who}},
\code{\link{pres84}},
\code{\link{pres88}},
\code{\link{pres92}},
\code{\link{vote84}},
\code{\link{vote88}},
\code{\link{vote96}}
}
\concept{Core}
\keyword{variable}
