% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_537.R
\name{voteelec}
\alias{voteelec}
\title{How important always to vote in elections}
\source{
General Social Survey https://gss.norc.org
}
\description{
voteelec
}
\details{
Question There are different opinions as to what it takes to be a good citizen. As far as you are concerned personally on a scale of 1 to 7, where 1 is not at all important and 7 is very important, how important is it…
A. Always to vote in elections
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all important
\item \verb{[7]} very important
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004, 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4907/vshow}{GSS Data Explorer page for \code{voteelec}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab not at all important \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab very important \tab can't choose \tab no answer \tab Total \cr
   2004 \tab 1340 \tab 41 \tab 22 \tab 35 \tab 79 \tab 144 \tab 161 \tab 975 \tab 12 \tab 3 \tab 2812 \cr
   2014 \tab 1274 \tab 53 \tab 21 \tab 40 \tab 77 \tab 119 \tab 155 \tab 772 \tab 23 \tab 4 \tab 2538 \cr
   Total \tab 2614 \tab 94 \tab 43 \tab 75 \tab 156 \tab 263 \tab 316 \tab 1747 \tab 35 \tab 7 \tab 5350 \cr
}
}

\concept{ISSP Citizenship}
\keyword{variable}
