% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_683.R
\name{voteprob}
\alias{voteprob}
\title{Voting problem in us elections}
\source{
General Social Survey https://gss.norc.org
}
\description{
voteprob
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} very strongly ineligible voters being allowed to vote
\item \code{2} moderately strongly ineligible voters being allowed to vote
\item \code{3} a little ineligible voters being allowed to vote
\item \code{4} a little eligible voters not being allowed to vote
\item \code{5} moderately strongly eligible voters not being allowed to vote
\item \code{6} very strongly eligible voters not being allowed to vote
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
