% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_684.R
\name{voteprob}
\alias{voteprob}
\title{Voting problem in us elections}
\source{
General Social Survey https://gss.norc.org
}
\description{
voteprob
}
\details{
Question If you had to choose, which of the following do you think is a bigger problem for U.S. elections? Ineligible voters being allowed to vote, or eligible voters not being allowed to vote?
}
\section{Values}{

\itemize{
\item \code{1} very strongly ineligible voters being allowed to vote
\item \code{2} moderately strongly ineligible voters being allowed to vote
\item \code{3} a little ineligible voters being allowed to vote
\item \code{4} a little eligible voters not being allowed to vote
\item \code{5} moderately strongly eligible voters not being allowed to vote
\item \code{6} very strongly eligible voters not being allowed to vote
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8191/vshow}{GSS Data Explorer page for \code{voteprob}}.
}

\seealso{
Other ANES: 
\code{\link{ratechall1ts}},
\code{\link{ratechall2ts}},
\code{\link{ratechall3ts}},
\code{\link{rateincumbts}},
\code{\link{whovotets}}
}
\concept{ANES}
\keyword{variable}
