% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_298.R
\name{wait10}
\alias{wait10}
\title{How long r wait treatment downhearted and b}
\source{
General Social Survey https://gss.norc.org
}
\description{
wait10
}
\details{
Question Given your life right now, how many days, weeks or months would you wait for this situation to get better on its own before seeking medical treatment?
J. Feeling downhearted and blue?
}
\section{Values}{

\itemize{
\item \verb{[0]} zero days
\item \verb{[991]} some days
\item \verb{[992]} some weeks
\item \verb{[993]} some months
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2708/vshow}{GSS Data Explorer page for \code{wait10}}.

Counts by year:\tabular{lllllllllllllllll}{
   year \tab iap \tab zero days \tab 1 \tab 10 \tab 12 \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 \tab some days \tab some weeks \tab some months \tab no answer \tab Total \cr
   2000 \tab 2146 \tab 8 \tab 149 \tab 1 \tab 3 \tab 114 \tab 25 \tab 8 \tab 4 \tab 6 \tab 4 \tab 18 \tab 33 \tab 25 \tab 273 \tab 2817 \cr
   Total \tab 2146 \tab 8 \tab 149 \tab 1 \tab 3 \tab 114 \tab 25 \tab 8 \tab 4 \tab 6 \tab 4 \tab 18 \tab 33 \tab 25 \tab 273 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{downcure}},
\code{\link{downfam}},
\code{\link{downqol}},
\code{\link{wait11}},
\code{\link{wait5}},
\code{\link{wait6}},
\code{\link{wait7}},
\code{\link{wait8}},
\code{\link{wait9}}
}
\concept{Health Status}
\keyword{variable}
