% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_298.R
\name{wait11}
\alias{wait11}
\title{How long r wait treatment health interfere}
\source{
General Social Survey https://gss.norc.org
}
\description{
wait11
}
\details{
Question Given your life right now, how many days, weeks or months would you wait for this situation to get better on its own before seeking medical treatment?
K. Having physical health or emotional problems interfere with you social life, like visiting friends, relatives, etc.?
}
\section{Values}{

\itemize{
\item \code{0} zero days
\item \code{991} some days
\item \code{992} some weeks
\item \code{993} some months
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2709/vshow}{GSS Data Explorer page for \code{wait11}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab iap \tab zero days \tab 1 \tab 11 \tab 12 \tab 14 \tab 2 \tab 3 \tab 30 \tab 4 \tab 5 \tab 6 \tab 7 \tab some days \tab some weeks \tab some months \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1922 \tab 9 \tab 265 \tab 1 \tab 3 \tab 1 \tab 179 \tab 33 \tab 1 \tab 16 \tab 7 \tab 9 \tab 6 \tab 23 \tab 48 \tab 25 \tab 3 \tab 266 \tab 2817 \cr
   Total \tab 1922 \tab 9 \tab 265 \tab 1 \tab 3 \tab 1 \tab 179 \tab 33 \tab 1 \tab 16 \tab 7 \tab 9 \tab 6 \tab 23 \tab 48 \tab 25 \tab 3 \tab 266 \tab 2817 \cr
}
}

\seealso{
Other Health Status: 
\code{\link{downcure}},
\code{\link{downfam}},
\code{\link{downqol}},
\code{\link{wait10}},
\code{\link{wait5}},
\code{\link{wait6}},
\code{\link{wait7}},
\code{\link{wait8}},
\code{\link{wait9}}
}
\concept{Health Status}
\keyword{variable}
