% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_298.R
\name{wait5}
\alias{wait5}
\title{How long r wait treatment doing less work e}
\source{
General Social Survey https://gss.norc.org
}
\description{
wait5
}
\details{
Question Given your life right now, how many days, weeks or months would you wait for this situation to get better on its own before seeking medical treatment?
E. Accomplishing less than you would like with your work or other regular daily activities as a result of any emotional problems (such as feeling depressed or anxious)?
}
\section{Values}{

\itemize{
\item \code{0} zero days
\item \code{991} some days
\item \code{992} some weeks
\item \code{993} some months
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2703/vshow}{GSS Data Explorer page for \code{wait5}}.
}

\seealso{
Other Health Status: 
\code{\link{downcure}},
\code{\link{downfam}},
\code{\link{downqol}},
\code{\link{wait10}},
\code{\link{wait11}},
\code{\link{wait6}},
\code{\link{wait7}},
\code{\link{wait8}},
\code{\link{wait9}}

Other Health: 
\code{\link{downcure}},
\code{\link{downfam}},
\code{\link{downqol}},
\code{\link{wait10}},
\code{\link{wait11}},
\code{\link{wait6}},
\code{\link{wait7}},
\code{\link{wait8}},
\code{\link{wait9}}
}
\concept{Health}
\concept{Health Status}
\keyword{variable}
