% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_503.R
\name{wantjob1}
\alias{wantjob1}
\title{Which job would you prefer}
\source{
General Social Survey https://gss.norc.org
}
\description{
wantjob1
}
\details{
Question Suppose you could decide on your work situation at present. Which of the following would you prefer?
}
\section{Values}{

\itemize{
\item \code{1} a full time job
\item \code{2} a part-time job
\item \code{3} a job with less than 10 hours a week
\item \code{4} no paid job at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4577/vshow}{GSS Data Explorer page for \code{wantjob1}}.
}

\seealso{
Other Work Orientation: 
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paytime}},
\code{\link{placewrk}},
\code{\link{sethours}},
\code{\link{techjobs}},
\code{\link{techwork}},
\code{\link{worknow}}
}
\concept{Work Orientation}
\keyword{variable}
