% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_505.R
\name{wantjob2}
\alias{wantjob2}
\title{Would like to have a job now or in the future}
\source{
General Social Survey https://gss.norc.org
}
\description{
wantjob2
}
\details{
Question Would you like to have a paid job, either now or in the future?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4596/vshow}{GSS Data Explorer page for \code{wantjob2}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1998 \tab 2460 \tab 158 \tab 180 \tab 1 \tab 33 \tab 2832 \cr
   2006 \tab 4008 \tab 277 \tab 220 \tab 2 \tab 3 \tab 4510 \cr
   2016 \tab 2326 \tab 303 \tab 207 \tab 8 \tab 23 \tab 2867 \cr
   Total \tab 8794 \tab 738 \tab 607 \tab 11 \tab 59 \tab 10209 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{evjob}},
\code{\link{getjob}},
\code{\link{leavejob}},
\code{\link{lookjob}},
\code{\link{priagncy}},
\code{\link{pubagncy}},
\code{\link{whyjbend}},
\code{\link{worryjob}},
\code{\link{yrjobend}}
}
\concept{Work Orientation}
\keyword{variable}
