% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_191.R
\name{warrntyc}
\alias{warrntyc}
\title{Did the seller provide a warranty?}
\source{
General Social Survey https://gss.norc.org
}
\description{
warrntyc
}
\details{
Question Did the seller guarantee to pay for some or all unexpected repairs needed for some period after the sale (i.e., did the seller provide a warranty)?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes -- informal assurance
\item \verb{[2]} yes -- written guarantee
\item \verb{[3]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1798/vshow}{GSS Data Explorer page for \code{warrntyc}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes -- informal assurance \tab yes -- written guarantee \tab no \tab don't know \tab no answer \tab Total \cr
   1996 \tab 2653 \tab 11 \tab 12 \tab 220 \tab 5 \tab 3 \tab 2904 \cr
   Total \tab 2653 \tab 11 \tab 12 \tab 220 \tab 5 \tab 3 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{cardealr}},
\code{\link{carprivt}},
\code{\link{newused}},
\code{\link{relprivt}},
\code{\link{typdealr}}
}
\concept{Markets}
\keyword{variable}
