% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_523.R
\name{waterfam}
\alias{waterfam}
\title{Water pollution danger to my family}
\source{
General Social Survey https://gss.norc.org
}
\description{
waterfam
}
\details{
Question B. And do you think that pollution of America's rivers, lakes, and streams is…
}
\section{Values}{

\itemize{
\item \code{1} extremely dangerous
\item \code{2} very dangerous
\item \code{3} somewhat dangerous
\item \code{4} not very dangerous
\item \code{5} not dangerous at all
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1993 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4763/vshow}{GSS Data Explorer page for \code{waterfam}}.
}

\seealso{
Other ISSP Environment: 
\code{\link{busdecid}},
\code{\link{chemfam}},
\code{\link{chemgen}},
\code{\link{indusfam}},
\code{\link{indusgen}},
\code{\link{pubdecid}},
\code{\link{tempfam}},
\code{\link{tempgen}},
\code{\link{watergen}}

Other Environment: 
\code{\link{busdecid}},
\code{\link{chemfam}},
\code{\link{chemgen}},
\code{\link{indusfam}},
\code{\link{indusgen}},
\code{\link{pubdecid}},
\code{\link{tempfam}},
\code{\link{tempgen}},
\code{\link{watergen}}
}
\concept{Environment}
\concept{ISSP Environment}
\keyword{variable}
