% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_210.R
\name{wayraise}
\alias{wayraise}
\title{How likely x s situation caused by the way raised}
\source{
General Social Survey https://gss.norc.org
}
\description{
wayraise
}
\details{
Question In your opinion, how likely is it that NAME's situation might be caused by (READ a-f) -- very likely, somewhat likely, not very likely, or not at all likely?
C. The way (s/he) was raised
}
\section{Values}{

\itemize{
\item \verb{[1]} very likely
\item \verb{[2]} somewhat likely
\item \verb{[3]} not very likely
\item \verb{[4]} not at all likely
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 2024 \cr
   A/B/C \tab 1996, 2006 \cr
   B/C/- \tab 2018 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1990/vshow}{GSS Data Explorer page for \code{wayraise}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab very likely \tab somewhat likely \tab not very likely \tab not at all likely \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 149 \tab 552 \tab 420 \tab 245 \tab 60 \tab 18 \tab 2904 \cr
   2006 \tab 3073 \tab 148 \tab 580 \tab 387 \tab 257 \tab 51 \tab 14 \tab 4510 \cr
   2018 \tab 1175 \tab 99 \tab 413 \tab 376 \tab 232 \tab 23 \tab 30 \tab 2348 \cr
   Total \tab 5708 \tab 396 \tab 1545 \tab 1183 \tab 734 \tab 134 \tab 62 \tab 9762 \cr
}
}

\seealso{
Other Mental Health: 
\code{\link{charactr}},
\code{\link{imbalnce}},
\code{\link{probfix3}},
\code{\link{probhlp1}},
\code{\link{probhlp2}},
\code{\link{probhlp3}},
\code{\link{seriousp}},
\code{\link{stresses}},
\code{\link{vigversn}}
}
\concept{Mental Health}
\keyword{variable}
