% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_391.R
\name{wealth}
\alias{wealth}
\title{Total wealth of respondent}
\source{
General Social Survey https://gss.norc.org
}
\description{
wealth
}
\details{
Question Please estimate your total wealth.
}
\section{Values}{

\itemize{
\item \code{1} less than $5,000
\item \code{2} $5,000 to $20,000
\item \code{3} $20,000 to $40,000
\item \code{4} $40,000 to $75,000
\item \code{5} $75,000 to $100,000
\item \code{6} $100,000 to $150,000
\item \code{7} $150,000 to $250,000
\item \code{8} $250,000 to $500,000
\item \code{9} $500,000 to $1 million
\item \code{10} $1 million to $2 million
\item \code{11} $2 million to $3 million
\item \code{12} $3 million to $4 million
\item \code{13} $4 million to $5 million
\item \code{14} $5 million to $10 million
\item \code{15} above $10 million
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2006 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3560/vshow}{GSS Data Explorer page for \code{wealth}}.
}

\seealso{
Other Shared Capitalism: 
\code{\link{compwage}},
\code{\link{empinput}},
\code{\link{emptrain}},
\code{\link{esop}},
\code{\link{extr2005}},
\code{\link{numorg}},
\code{\link{ops2005}},
\code{\link{slfmangd}},
\code{\link{talkteam}}
}
\concept{Shared Capitalism}
\keyword{variable}
