% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_289.R
\name{webable}
\alias{webable}
\title{R's ability to use internet}
\source{
General Social Survey https://gss.norc.org
}
\description{
webable
}
\details{
Question Would you rate your ability to use the World Wide Web as excellent, good, fair, poor, or very poor?
}
\section{Values}{

\itemize{
\item \code{1} excellent
\item \code{2} good
\item \code{3} fair
\item \code{4} poor
\item \code{5} very poor
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2621/vshow}{GSS Data Explorer page for \code{webable}}.
}

\seealso{
Other Information Society: 
\code{\link{emailyr}},
\code{\link{jobcntct}},
\code{\link{jobcowrk}},
\code{\link{jobfrnds}},
\code{\link{jobonwww}},
\code{\link{jobplace}},
\code{\link{jobpub}},
\code{\link{jobtvrad}},
\code{\link{webyr}}

Other Computers And Internet: 
\code{\link{emailyr}},
\code{\link{webyr}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
