% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_289.R
\name{webable}
\alias{webable}
\title{R's ability to use internet}
\source{
General Social Survey https://gss.norc.org
}
\description{
webable
}
\details{
Question Would you rate your ability to use the World Wide Web as excellent, good, fair, poor, or very poor?
}
\section{Values}{

\itemize{
\item \code{1} excellent
\item \code{2} good
\item \code{3} fair
\item \code{4} poor
\item \code{5} very poor
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002 \cr
   B/C/- \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2621/vshow}{GSS Data Explorer page for \code{webable}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab excellent \tab good \tab fair \tab poor \tab very poor \tab no answer \tab Total \cr
   2002 \tab 1914 \tab 218 \tab 311 \tab 197 \tab 45 \tab 15 \tab 65 \tab 2765 \cr
   2004 \tab 2122 \tab 192 \tab 289 \tab 139 \tab 37 \tab 21 \tab 12 \tab 2812 \cr
   Total \tab 4036 \tab 410 \tab 600 \tab 336 \tab 82 \tab 36 \tab 77 \tab 5577 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{emailyr}},
\code{\link{jobcntct}},
\code{\link{jobcowrk}},
\code{\link{jobfrnds}},
\code{\link{jobonwww}},
\code{\link{jobplace}},
\code{\link{jobpub}},
\code{\link{jobtvrad}},
\code{\link{webyr}}
}
\concept{Information Society}
\keyword{variable}
