% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_648.R
\name{webrely}
\alias{webrely}
\title{Hard to distinguish reliable and unreliable info online}
\source{
General Social Survey https://gss.norc.org
}
\description{
webrely
}
\details{
Question (To what extent do you agree or disagree with the following statements?) It is not easy to distinguish between reliable and unreliable health information on the internet.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7817/vshow}{GSS Data Explorer page for \code{webrely}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2381 \tab 163 \tab 476 \tab 268 \tab 159 \tab 20 \tab 60 \tab 9 \tab 8 \tab 3544 \cr
   Total \tab 2381 \tab 163 \tab 476 \tab 268 \tab 159 \tab 20 \tab 60 \tab 9 \tab 8 \tab 3544 \cr
}
}

\keyword{variable}
