% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_262.R
\name{webtv}
\alias{webtv}
\title{R has internet via webtv}
\source{
General Social Survey https://gss.norc.org
}
\description{
webtv
}
\details{
Question A. Do you have access to the Internet or World Wide Web in your home through Web TV?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2379/vshow}{GSS Data Explorer page for \code{webtv}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   2000 \tab 1814 \tab 23 \tab 970 \tab 10 \tab - \tab 2817 \cr
   2002 \tab 1835 \tab 53 \tab 860 \tab 6 \tab 11 \tab 2765 \cr
   Total \tab 3649 \tab 76 \tab 1830 \tab 16 \tab 11 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{compuse}},
\code{\link{webmob}}
}
\concept{Information Society}
\keyword{variable}
