% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_289.R
\name{webyr}
\alias{webyr}
\title{When r first began using web}
\source{
General Social Survey https://gss.norc.org
}
\description{
webyr
}
\details{
Question B. When did you first begin using the World Wide Web (for purposes other than email)?
}
\section{Values}{

\itemize{
\item \code{1} this year
\item \code{2} in the last year
\item \code{3} 1-2 years ago
\item \code{4} 2-3 years ago
\item \code{5} 3-4 years ago
\item \code{6} 4-5 years ago
\item \code{7} 5 or more years ago
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2620/vshow}{GSS Data Explorer page for \code{webyr}}.
}

\seealso{
Other Information Society: 
\code{\link{emailyr}},
\code{\link{jobcntct}},
\code{\link{jobcowrk}},
\code{\link{jobfrnds}},
\code{\link{jobonwww}},
\code{\link{jobplace}},
\code{\link{jobpub}},
\code{\link{jobtvrad}},
\code{\link{webable}}

Other Computers And Internet: 
\code{\link{emailyr}},
\code{\link{webable}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
