% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_600.R
\name{welfare1}
\alias{welfare1}
\title{Welfare makes people work less}
\source{
General Social Survey https://gss.norc.org
}
\description{
welfare1
}
\details{
Question Here are some opinions other people have expressed about welfare. For each of the following statements, please tell me whether you strongly agree, agree, disagree, or strongly disagree with it.
Welfare . . . (READ EACH STATEMENT. CIRCLE ONE CODE FOR EACH.)
A. Makes people work less than they would if there wasn''t a welfare system.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1986 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5498/vshow}{GSS Data Explorer page for \code{welfare1}}.
}

\seealso{
Other Welfare: 
\code{\link{welfare2}},
\code{\link{welfare3}}

Other Income: 
\code{\link{welfare2}},
\code{\link{welfare3}}

Other Welfare: 
\code{\link{welfare2}},
\code{\link{welfare3}}
}
\concept{Income}
\concept{Welfare}
\keyword{variable}
