% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_664.R
\name{whatco2}
\alias{whatco2}
\title{Classify place where partners works}
\source{
General Social Survey https://gss.norc.org
}
\description{
whatco2
}
\details{
Question How would you classify the place where your (husband/wife/spouse/partner) (worked/works) Was this mainly….?
}
\section{Values}{

\itemize{
\item \verb{[1]} manufacturing
\item \verb{[2]} wholesale trade
\item \verb{[3]} retail trade
\item \verb{[4]} other (agriculture, construction, service, government, etc.)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8012/vshow}{GSS Data Explorer page for \code{whatco2}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab manufacturing \tab wholesale trade \tab retail trade \tab other (agriculture, construction, service, government, etc.) \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 3244 \tab 41 \tab 6 \tab 45 \tab 195 \tab 7 \tab 3 \tab 3 \tab 3544 \cr
   Total \tab 7276 \tab 41 \tab 6 \tab 45 \tab 195 \tab 7 \tab 3 \tab 3 \tab 7576 \cr
}
}

\keyword{variable}
