% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_641.R
\name{whatma2}
\alias{whatma2}
\title{How would you classify the place where mother (or youngest same-sex parent) worked? Was this mainly...?}
\source{
General Social Survey https://gss.norc.org
}
\description{
whatma2
}
\details{
Question How would you classify the place where your mother worked? Was this mainly….?
}
\section{Values}{

\itemize{
\item \code{1} manufacturing
\item \code{2} wholesale trade
\item \code{3} retail trade
\item \code{4} other (agriculture, construction, service, government, etc.)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7745/vshow}{GSS Data Explorer page for \code{whatma2}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab manufacturing \tab wholesale trade \tab retail trade \tab other (agriculture, construction, service, government, etc.) \tab 5 \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 934 \tab 276 \tab 44 \tab 402 \tab 1718 \tab 162 \tab 2 \tab 6 \tab 3544 \cr
   Total \tab 934 \tab 276 \tab 44 \tab 402 \tab 1718 \tab 162 \tab 2 \tab 6 \tab 3544 \cr
}
}

\keyword{variable}
