% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_642.R
\name{whatpa2}
\alias{whatpa2}
\title{How would you classify the place where father (or oldest same-sex parent) worked? Was this mainly...?}
\source{
General Social Survey https://gss.norc.org
}
\description{
whatpa2
}
\details{
Question How would you classify the place where your father worked? Was this mainly….?
}
\section{Values}{

\itemize{
\item \code{1} manufacturing
\item \code{2} wholesale trade
\item \code{3} retail trade
\item \code{4} other (agriculture, construction, service, government, etc.)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7744/vshow}{GSS Data Explorer page for \code{whatpa2}}.
}

\keyword{variable}
