% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_641.R
\name{whatslf2}
\alias{whatslf2}
\title{How would you classify the place where you (#whatslf2fill)? Was this mainly...?}
\source{
General Social Survey https://gss.norc.org
}
\description{
whatslf2
}
\details{
Question How would you classify the place where you (work/worked)? Was this mainly...?
}
\section{Values}{

\itemize{
\item \code{1} manufacturing
\item \code{2} wholesale trade
\item \code{3} retail trade
\item \code{4} other (agriculture, construction, service, government, etc.)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7746/vshow}{GSS Data Explorer page for \code{whatslf2}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab manufacturing \tab wholesale trade \tab retail trade \tab other (agriculture, construction, service, government, etc.) \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 1326 \tab 228 \tab 62 \tab 283 \tab 1612 \tab 16 \tab 12 \tab 5 \tab 3544 \cr
   Total \tab 1326 \tab 228 \tab 62 \tab 283 \tab 1612 \tab 16 \tab 12 \tab 5 \tab 3544 \cr
}
}

\keyword{variable}
