% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_641.R
\name{whatsp2}
\alias{whatsp2}
\title{How would you classify the place where your (#_spfill1) (#whatsp2fill)? Was this mainly...?}
\source{
General Social Survey https://gss.norc.org
}
\description{
whatsp2
}
\details{
Question How would you classify the place where your (spouse/partner) (work/worked)? Was this mainly...?
}
\section{Values}{

\itemize{
\item \verb{[1]} manufacturing
\item \verb{[2]} wholesale trade
\item \verb{[3]} retail trade
\item \verb{[4]} other (agriculture, construction, service, government, etc.)
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7748/vshow}{GSS Data Explorer page for \code{whatsp2}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab manufacturing \tab wholesale trade \tab retail trade \tab other (agriculture, construction, service, government, etc.) \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 4032 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 4032 \cr
   2022 \tab 2132 \tab 163 \tab 36 \tab 148 \tab 1019 \tab 31 \tab 11 \tab 4 \tab 3544 \cr
   Total \tab 6164 \tab 163 \tab 36 \tab 148 \tab 1019 \tab 31 \tab 11 \tab 4 \tab 7576 \cr
}
}

\keyword{variable}
