% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_146.R
\name{wherewrk}
\alias{wherewrk}
\title{Doing your job in one or several different places}
\source{
General Social Survey https://gss.norc.org
}
\description{
wherewrk
}
\details{
Question When you are doing your job, do you usually do it in one place (like the same street address) or in several different places?
}
\section{Values}{

\itemize{
\item \verb{[1]} some place
\item \verb{[2]} different
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1991 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1372/vshow}{GSS Data Explorer page for \code{wherewrk}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab some place \tab different \tab no answer \tab Total \cr
   1991 \tab 605 \tab 626 \tab 279 \tab 7 \tab 1517 \cr
   Total \tab 605 \tab 626 \tab 279 \tab 7 \tab 1517 \cr
}
}

\seealso{
Other Work Organizations: 
\code{\link{baseofop}},
\code{\link{exptrain}},
\code{\link{findout}},
\code{\link{learning}},
\code{\link{splocnum}},
\code{\link{training}}
}
\concept{Work Organizations}
\keyword{variable}
