% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_185.R
\name{whnangry}
\alias{whnangry}
\title{How long ago did this event happen?}
\source{
General Social Survey https://gss.norc.org
}
\description{
whnangry
}
\details{
Question How long ago did this event happen? (ALLOW OPEN-ENDED RESPONSE; IF A RANGE IS GIVEN, PROBE FOR SPECIFIC NUMBER)
}
\section{Values}{

\itemize{
\item \code{1} today
\item \code{2} yesterday
\item \code{3} up to a week ago
\item \code{4} up to a month ago
\item \code{5} months ago
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1740/vshow}{GSS Data Explorer page for \code{whnangry}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab today \tab yesterday \tab up to a week ago \tab up to a month ago \tab months ago \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1779 \tab 79 \tab 108 \tab 500 \tab 350 \tab 78 \tab 1 \tab 9 \tab 2904 \cr
   Total \tab 1779 \tab 79 \tab 108 \tab 500 \tab 350 \tab 78 \tab 1 \tab 9 \tab 2904 \cr
}
}

\seealso{
Other Emotions: 
\code{\link{angryev}},
\code{\link{angryfam}},
\code{\link{angrygvt}},
\code{\link{angrywhy}},
\code{\link{angrywrk}},
\code{\link{beplesnt}},
\code{\link{madat1}},
\code{\link{notupset}},
\code{\link{notworry}}
}
\concept{Emotions}
\keyword{variable}
