% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_606.R
\name{whoelse5}
\alias{whoelse5}
\title{Presence of others: other adults}
\source{
General Social Survey https://gss.norc.org
}
\description{
whoelse5
}
\details{
Question What other persons were present during the interview? Circle all that apply.
E. Other Adults
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C/- \tab 1994, 1996, 1998, 2000, 2002, 2004, 2008, 2010, 2012, 2014, 2016, 2018, 2022 \cr
   A/B/C/D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5554/vshow}{GSS Data Explorer page for \code{whoelse5}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab not available in this year \tab Total \cr
   1994 \tab 1518 \tab 77 \tab 1354 \tab 8 \tab 35 \tab - \tab 2992 \cr
   1996 \tab - \tab 214 \tab 2612 \tab 7 \tab 71 \tab - \tab 2904 \cr
   1998 \tab - \tab 204 \tab 2507 \tab 4 \tab 117 \tab - \tab 2832 \cr
   2000 \tab - \tab 178 \tab 2491 \tab 2 \tab 146 \tab - \tab 2817 \cr
   2002 \tab - \tab 149 \tab 2601 \tab 15 \tab - \tab - \tab 2765 \cr
   2004 \tab - \tab 148 \tab 2631 \tab 21 \tab 12 \tab - \tab 2812 \cr
   2006 \tab - \tab 228 \tab 4258 \tab 13 \tab 11 \tab - \tab 4510 \cr
   2008 \tab - \tab 102 \tab 1918 \tab - \tab 3 \tab - \tab 2023 \cr
   2010 \tab - \tab 181 \tab 1855 \tab - \tab 8 \tab - \tab 2044 \cr
   2012 \tab - \tab 122 \tab 1849 \tab - \tab 3 \tab - \tab 1974 \cr
   2014 \tab - \tab 197 \tab 2335 \tab - \tab 6 \tab - \tab 2538 \cr
   2016 \tab - \tab 219 \tab 2641 \tab - \tab 7 \tab - \tab 2867 \cr
   2018 \tab 9 \tab 198 \tab 2141 \tab - \tab - \tab - \tab 2348 \cr
   2022 \tab 1642 \tab 136 \tab 1672 \tab 94 \tab - \tab - \tab 3544 \cr
   Total \tab 3169 \tab 2353 \tab 32865 \tab 164 \tab 419 \tab 0 \tab 38970 \cr
}
}

\seealso{
Other Paradata: 
\code{\link{feelevel}},
\code{\link{feeused}},
\code{\link{intid}},
\code{\link{saqgene}},
\code{\link{saqissp}},
\code{\link{saqsex}},
\code{\link{whoelse3}},
\code{\link{whoelse4}},
\code{\link{whoelse6}}
}
\concept{Paradata}
\keyword{variable}
