% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_196.R
\name{wholoand}
\alias{wholoand}
\title{From whom did you borrow this money?}
\source{
General Social Survey https://gss.norc.org
}
\description{
wholoand
}
\details{
Question Think of the most recent loan of this type. From whom did you borrow this money?
}
\section{Values}{

\itemize{
\item \code{1} parents or grandparents
\item \code{2} children
\item \code{3} other relative
\item \code{4} friend
\item \code{5} co-worker or bus. associate
\item \code{6} other-specify
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1850/vshow}{GSS Data Explorer page for \code{wholoand}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab parents or grandparents \tab children \tab other relative \tab friend \tab co-worker or bus. associate \tab other-specify \tab no answer \tab Total \cr
   1996 \tab 2746 \tab 68 \tab 4 \tab 37 \tab 35 \tab 5 \tab 5 \tab 4 \tab 2904 \cr
   Total \tab 2746 \tab 68 \tab 4 \tab 37 \tab 35 \tab 5 \tab 5 \tab 4 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{borrowed}},
\code{\link{fixer1}},
\code{\link{fixerc}},
\code{\link{fixerd}},
\code{\link{fixere}},
\code{\link{fixerf}},
\code{\link{relfixer}},
\code{\link{satfixes}},
\code{\link{sellbed}}
}
\concept{Markets}
\keyword{variable}
