% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_174.R
\name{whoteach}
\alias{whoteach}
\title{Should only minorities teach minority history}
\source{
General Social Survey https://gss.norc.org
}
\description{
whoteach
}
\details{
Question Some people say that when the experience of a racial or ethnic minority group is taught in high school or college, it should be taught by a teacher from that group. Do you strongly agree, agree, neither agree nor disagree, disagree, or strongly disagree?
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1631/vshow}{GSS Data Explorer page for \code{whoteach}}.
}

\seealso{
Other Multiculturalism: 
\code{\link{colaff}},
\code{\link{colaffy}},
\code{\link{congeth}},
\code{\link{discaffy}},
\code{\link{ethhist}},
\code{\link{jobaff}},
\code{\link{owneth}},
\code{\link{schleth}},
\code{\link{teacheth}}
}
\concept{Multiculturalism}
\keyword{variable}
