% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_683.R
\name{whovote24}
\alias{whovote24}
\title{Voting intention in 2024 election (pre july 25 2024)}
\source{
General Social Survey https://gss.norc.org
}
\description{
whovote24
}
\details{
Question In the November election for President, do you think you will probably vote for Donald Trump, Joe Biden, someone else, or will you probably not vote?
}
\section{Values}{

\itemize{
\item \code{1} donald trump
\item \code{2} joe biden
\item \code{3} someone else
\item \code{4} probably not vote
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8193/vshow}{GSS Data Explorer page for \code{whovote24}}.
}

\seealso{
Other ANES: 
\code{\link{econstat}},
\code{\link{lkelyvot}},
\code{\link{polintrst}},
\code{\link{ratechall124}},
\code{\link{ratechall1ts}},
\code{\link{ratechall224}},
\code{\link{ratechall2ts}},
\code{\link{ratechall324}},
\code{\link{ratechall3ts}},
\code{\link{ratedemp}},
\code{\link{rateincumb24}},
\code{\link{rateincumbts}},
\code{\link{ratemaga}},
\code{\link{raterepp}},
\code{\link{satdemoc}},
\code{\link{voteprob}},
\code{\link{whovote24a}},
\code{\link{whovotets}}
}
\concept{ANES}
\keyword{variable}
