% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_671.R
\name{whovote24a}
\alias{whovote24a}
\title{Voting intention in 2024 election (post july 25 2024)}
\source{
General Social Survey https://gss.norc.org
}
\description{
whovote24a
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} donald trump
\item \code{2} kamala harris
\item \code{3} someone else
\item \code{4} probably not vote
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
