% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_684.R
\name{whovotets}
\alias{whovotets}
\title{Voting intention - time series}
\source{
General Social Survey https://gss.norc.org
}
\description{
whovotets
}
\details{
Question (Recode of WHOVOTE24/WHOVOTE24A)
}
\section{Values}{

\itemize{
\item \code{1} republican candidate
\item \code{2} democratic candidate
\item \code{3} someone else
\item \code{4} probably not vote
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2024 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/8194/vshow}{GSS Data Explorer page for \code{whovotets}}.
}

\seealso{
Other ANES: 
\code{\link{ratechall1ts}},
\code{\link{ratechall2ts}},
\code{\link{ratechall3ts}},
\code{\link{rateincumbts}},
\code{\link{voteprob}}
}
\concept{ANES}
\keyword{variable}
