% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_258.R
\name{whtcom}
\alias{whtcom}
\title{Does r know whites from community}
\source{
General Social Survey https://gss.norc.org
}
\description{
whtcom
}
\details{
Question Now, I’m going to ask some questions about these people you personally know.
B. Do you know any of these Whites from the community where you now live?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2336/vshow}{GSS Data Explorer page for \code{whtcom}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2485 \tab 232 \tab 63 \tab 37 \tab 2817 \cr
   Total \tab 2485 \tab 232 \tab 63 \tab 37 \tab 2817 \cr
}
}

\seealso{
Other Multi-Ethnic United States: 
\code{\link{blkschl}},
\code{\link{knwasn}},
\code{\link{knwblk}},
\code{\link{knwhisp}},
\code{\link{knwjew}},
\code{\link{whtcls}},
\code{\link{whtrel}},
\code{\link{whtschl}},
\code{\link{whtwrk}}
}
\concept{Multi-Ethnic United States}
\keyword{variable}
