% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_175.R
\name{whtgovt}
\alias{whtgovt}
\title{Do whites get proper govt attention}
\source{
General Social Survey https://gss.norc.org
}
\description{
whtgovt
}
\details{
Question A. Do you think that white people in this country get much more attention from government than they deserve, more attention than they deserve, about the right amount of attention, less attention than they deserve, or much less attention from government than they deserve?
}
\section{Values}{

\itemize{
\item \code{1} much more
\item \code{2} more
\item \code{3} about right
\item \code{4} less
\item \code{5} much less
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1994 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1632/vshow}{GSS Data Explorer page for \code{whtgovt}}.
}

\seealso{
Other Multiculturalism: 
\code{\link{asngovt}},
\code{\link{blkgovt}},
\code{\link{econpast}},
\code{\link{hspgovt}},
\code{\link{obrespct}},
\code{\link{wlthimm}},
\code{\link{wlthundc}},
\code{\link{workimm}},
\code{\link{workundc}}

Other Ethnicity: 
\code{\link{asngovt}},
\code{\link{blkgovt}},
\code{\link{hspgovt}}

Other Federal Government: 
\code{\link{asngovt}},
\code{\link{blkgovt}},
\code{\link{hspgovt}}

Other Race: 
\code{\link{asngovt}},
\code{\link{blkgovt}},
\code{\link{hspgovt}}
}
\concept{Ethnicity}
\concept{Federal Government}
\concept{Multiculturalism}
\concept{Race}
\keyword{variable}
