% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_494.R
\name{whydisc5}
\alias{whydisc5}
\title{Why were you discriminated against}
\source{
General Social Survey https://gss.norc.org
}
\description{
whydisc5
}
\details{
Question (HAND CARD D4)
In your opinion, what was the main reason for the discrimination?
Please select one reason.
}
\section{Values}{

\itemize{
\item \code{1} my age
\item \code{2} my race, ethnicity
\item \code{3} my nationality
\item \code{4} my sex
\item \code{5} my religion
\item \code{6} my disability/mental or physical illness
\item \code{7} my family responsibilities
\item \code{8} my political beliefs
\item \code{9} other reason
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6326/vshow}{GSS Data Explorer page for \code{whydisc5}}.

Counts by year:\tabular{llllllllllllll}{
   year \tab iap \tab my age \tab my race, ethnicity \tab my nationality \tab my sex \tab my religion \tab my disability/mental or physical illness \tab my family responsibilities \tab my political beliefs \tab other reason \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2595 \tab 63 \tab 56 \tab 10 \tab 36 \tab 2 \tab 8 \tab 15 \tab 3 \tab 73 \tab 3 \tab 3 \tab 2867 \cr
   Total \tab 2595 \tab 63 \tab 56 \tab 10 \tab 36 \tab 2 \tab 8 \tab 15 \tab 3 \tab 73 \tab 3 \tab 3 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{discwk5}},
\code{\link{famorjob}},
\code{\link{flexhrs}},
\code{\link{flexhrs1}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{joborfam}},
\code{\link{wkpersnl}},
\code{\link{wrkindp}}
}
\concept{Work Orientation}
\keyword{variable}
