% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_424.R
\name{whyjbct2}
\alias{whyjbct2}
\title{Why job abolished or laid off}
\source{
General Social Survey https://gss.norc.org
}
\description{
whyjbct2
}
\details{
Question Why was the job abolished? Why were you permanently laid off?
}
\section{Values}{

\itemize{
\item \code{1} company's sales were declining
\item \code{2} machines and computers made my job obsolete
\item \code{3} job requirements changed and i was not able to do the job
\item \code{4} job moved to another location in the united states
\item \code{5} job moved to another country
\item \code{6} other
\item \code{7} company sold
\item \code{8} job done
\item \code{9} company in financial problem
\item \code{10} end of contract
\item \code{11} seasonal work
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3878/vshow}{GSS Data Explorer page for \code{whyjbct2}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab company's sales were declining \tab job requirements changed and i was not able to do the job \tab job moved to another country \tab Total \cr
   2008 \tab 2016 \tab 5 \tab 1 \tab 1 \tab 2023 \cr
   Total \tab 2016 \tab 5 \tab 1 \tab 1 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{findnwjb}},
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{moretrde}},
\code{\link{newjobs}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}},
\code{\link{wkcomptr}}
}
\concept{Global Economics}
\keyword{variable}
