% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_423.R
\name{whyleave}
\alias{whyleave}
\title{Why r left last job}
\source{
General Social Survey https://gss.norc.org
}
\description{
whyleave
}
\details{
Question Why did you leave your last (full-time) job? Did you leave voluntarily, was the job abolished, or were you permanently laid off?
}
\section{Values}{

\itemize{
\item \code{1} left voluntarily
\item \code{2} job was abolished
\item \code{3} permanently laid off
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3876/vshow}{GSS Data Explorer page for \code{whyleave}}.
}

\seealso{
Other Global Economics: 
\code{\link{curempyr}},
\code{\link{pastpay}},
\code{\link{paychnge}},
\code{\link{whyjbct1}},
\code{\link{work3yrs}}

Other Employment: 
\code{\link{curempyr}},
\code{\link{whyjbct1}},
\code{\link{work3yrs}}

Other Job: 
\code{\link{curempyr}},
\code{\link{pastpay}},
\code{\link{paychnge}},
\code{\link{whyjbct1}},
\code{\link{work3yrs}}

Other Unemployment: 
\code{\link{whyjbct1}}
}
\concept{Employment}
\concept{Global Economics}
\concept{Job}
\concept{Unemployment}
\keyword{variable}
