% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_424.R
\name{whylose2}
\alias{whylose2}
\title{Why that much possibility of job abolished or laid off}
\source{
General Social Survey https://gss.norc.org
}
\description{
whylose2
}
\details{
Question B. (Reason 2)
}
\section{Values}{

\itemize{
\item \code{1} company's sales were declining
\item \code{2} machines and computers made my job obsolete
\item \code{3} job requirements changed and i was not able to do the job
\item \code{4} job moved to another location in the united states
\item \code{5} job moved to another country
\item \code{6} other
\item \code{7} company sold
\item \code{8} job done
\item \code{9} company in financial problem
\item \code{10} end of contract
\item \code{11} seasonal work
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3881/vshow}{GSS Data Explorer page for \code{whylose2}}.
}

\seealso{
Other Global Economics: 
\code{\link{agecmeus}},
\code{\link{autonojb}},
\code{\link{bornsp}},
\code{\link{curempyr}},
\code{\link{emphlth}},
\code{\link{emphplan}},
\code{\link{findnwjb}},
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{moretrde}},
\code{\link{natdeg}},
\code{\link{newjobs}},
\code{\link{othplan}},
\code{\link{pastpay}},
\code{\link{paychnge}},
\code{\link{typevisa}},
\code{\link{visa}},
\code{\link{whyjbct1}},
\code{\link{whyjbct2}},
\code{\link{whyleave}},
\code{\link{whylose1}},
\code{\link{whylose3}},
\code{\link{whynojb1}},
\code{\link{whynojb2}},
\code{\link{wkcomptr}},
\code{\link{wocomptr}},
\code{\link{work3yrs}}
}
\concept{Global Economics}
\keyword{variable}
