% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_425.R
\name{whynojb2}
\alias{whynojb2}
\title{Why job abolished or laid off}
\source{
General Social Survey https://gss.norc.org
}
\description{
whynojb2
}
\details{
Question Why job abolished or laid off.
}
\section{Values}{

\itemize{
\item \code{1} company's sales were declining
\item \code{2} machines and computers made my job obsolete
\item \code{3} job requirements changed and i was not able to do the job
\item \code{4} job moved to another location in the united states
\item \code{5} job moved to another country
\item \code{6} other
\item \code{7} company sold
\item \code{8} job done
\item \code{9} company in financial problem
\item \code{10} end of contract
\item \code{11} seasonal work
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3896/vshow}{GSS Data Explorer page for \code{whynojb2}}.
}

\seealso{
Other Global Economics: 
\code{\link{autonojb}},
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{whynojb1}},
\code{\link{wocomptr}}
}
\concept{Global Economics}
\keyword{variable}
