% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_126.R
\name{whypoor2}
\alias{whypoor2}
\title{Loose morals and drunkenness}
\source{
General Social Survey https://gss.norc.org
}
\description{
whypoor2
}
\details{
Question Now I will a list of reasons some people give to explain why there are poor people in this country. Please tell me whether you feel each of these is very important, somewhat important, or not important in explaining why there are poor people in this country.
B. Loose morals and drunkenness
}
\section{Values}{

\itemize{
\item \code{1} very important
\item \code{2} somewhat important
\item \code{3} not important
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1171/vshow}{GSS Data Explorer page for \code{whypoor2}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab very important \tab somewhat important \tab not important \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 521 \tab 461 \tab 338 \tab 44 \tab 8 \tab 1372 \cr
   Total \tab 0 \tab 521 \tab 461 \tab 338 \tab 44 \tab 8 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{socdif1}},
\code{\link{socdif2}},
\code{\link{socdif3}},
\code{\link{socdif4}},
\code{\link{whypoor1}},
\code{\link{whypoor3}},
\code{\link{whypoor4}},
\code{\link{wlthjews}},
\code{\link{wlthwhts}}
}
\concept{Intergroup Relations}
\keyword{variable}
