% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_305.R
\name{whywkhme}
\alias{whywkhme}
\title{Usual reason r work at home}
\source{
General Social Survey https://gss.norc.org
}
\description{
whywkhme
}
\details{
Question Is it usually because you want to, you have to in order to keep up with your job, or for some other reason?
}
\section{Values}{

\itemize{
\item \verb{[1]} worker wants to work at home
\item \verb{[2]} worker has to work at home to keep up with job
\item \verb{[3]} worker is operating home-based business
\item \verb{[4]} other combinations and other reasons
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2768/vshow}{GSS Data Explorer page for \code{whywkhme}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab worker wants to work at home \tab worker has to work at home to keep up with job \tab worker is operating home-based business \tab no answer \tab don't know \tab other combinations and other reasons \tab skipped on web \tab Total \cr
   2002 \tab 2077 \tab 254 \tab 224 \tab 197 \tab 13 \tab - \tab - \tab - \tab 2765 \cr
   2006 \tab 3803 \tab 257 \tab 225 \tab 213 \tab 11 \tab 1 \tab - \tab - \tab 4510 \cr
   2010 \tab 1531 \tab 160 \tab 189 \tab 60 \tab 27 \tab 3 \tab 74 \tab - \tab 2044 \cr
   2014 \tab 2033 \tab 142 \tab 203 \tab 84 \tab 7 \tab - \tab 69 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 171 \tab 200 \tab 76 \tab 896 \tab 3 \tab 102 \tab - \tab 2348 \cr
   2022 \tab 2641 \tab 476 \tab 160 \tab 106 \tab 3 \tab 8 \tab 130 \tab 20 \tab 3544 \cr
   Total \tab 12985 \tab 1460 \tab 1201 \tab 736 \tab 957 \tab 15 \tab 375 \tab 20 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{famvswk}},
\code{\link{famwkoff}},
\code{\link{hrsrelax}},
\code{\link{learnnew}},
\code{\link{secondwk}},
\code{\link{wkvsfam}},
\code{\link{workdiff}},
\code{\link{workfast}},
\code{\link{wrkhome}}
}
\concept{Quality of Working Life}
\keyword{variable}
