% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_305.R
\name{whywkhme}
\alias{whywkhme}
\title{Usual reason r work at home}
\source{
General Social Survey https://gss.norc.org
}
\description{
whywkhme
}
\details{
Question Is it usually because you want to, you have to in order to keep up with your job, or for some other reason?
}
\section{Values}{

\itemize{
\item \code{1} worker wants to work at home
\item \code{2} worker has to work at home to keep up with job
\item \code{3} worker is operating home-based business
\item \code{4} other combinations and other reasons
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2768/vshow}{GSS Data Explorer page for \code{whywkhme}}.
}

\seealso{
Other Quality of Working Life: 
\code{\link{famvswk}},
\code{\link{famwkoff}},
\code{\link{hrsrelax}},
\code{\link{learnnew}},
\code{\link{secondwk}},
\code{\link{wkvsfam}},
\code{\link{workdiff}},
\code{\link{workfast}},
\code{\link{wrkhome}}

Other Employment: 
\code{\link{wrkhome}}

Other Work: 
\code{\link{famwkoff}},
\code{\link{wrkhome}}
}
\concept{Employment}
\concept{Quality of Working Life}
\concept{Work}
\keyword{variable}
