% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_177.R
\name{willwed1}
\alias{willwed1}
\title{How likely you two will get married?}
\source{
General Social Survey https://gss.norc.org
}
\description{
willwed1
}
\details{
Question How likely is it that you two will get married? Would you say that it is very likely, somewhat likely, somewhat unlikely, or very unlikely?
}
\section{Values}{

\itemize{
\item \code{1} very likely
\item \code{2} somewhat likely
\item \code{3} somewhat unlikely
\item \code{4} very unlikely
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1652/vshow}{GSS Data Explorer page for \code{willwed1}}.
}

\seealso{
Other Gender: 
\code{\link{emoteoth}},
\code{\link{fejobaff}},
\code{\link{hapboys}},
\code{\link{hapgirls}},
\code{\link{rhmewrk}},
\code{\link{sharesep}},
\code{\link{sphmewrk}},
\code{\link{tradmod}},
\code{\link{willwed2}}

Other Marriage: 
\code{\link{emoteoth}},
\code{\link{sharesep}},
\code{\link{tradmod}},
\code{\link{willwed2}}

Other Romantic Relationships: 
\code{\link{willwed2}}
}
\concept{Gender}
\concept{Marriage}
\concept{Romantic Relationships}
\keyword{variable}
