% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_34.R
\name{wirtap}
\alias{wirtap}
\title{Wiretapping}
\source{
General Social Survey https://gss.norc.org
}
\description{
wirtap
}
\details{
Question Everything considered, would you say that, in general, you approve or disapprove of wiretapping?
}
\section{Values}{

\itemize{
\item \code{1} approve
\item \code{2} disapprove
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1974 \tab -/- \tab full \cr
   1975 \tab -/- \tab full \cr
   1977 \tab -/- \tab full \cr
   1978 \tab -/- \tab full \cr
   1982 \tab -/- \tab full \cr
   1983 \tab -/- \tab full \cr
   1985 \tab -/- \tab full \cr
   1986 \tab -/- \tab full \cr
   1988 \tab A/B \tab partial \cr
   1989 \tab A/B \tab partial \cr
   1990 \tab A/B \tab partial \cr
   1991 \tab A/B \tab partial \cr
   1993 \tab A/B \tab partial \cr
   1994 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/284/vshow}{GSS Data Explorer page for \code{wirtap}}.
}

\seealso{
Other Core: 
\code{\link{denom}},
\code{\link{grass}},
\code{\link{grassy}},
\code{\link{lawfirm}},
\code{\link{lawimp}},
\code{\link{lawinfo}},
\code{\link{other}},
\code{\link{othjew}},
\code{\link{relig}}
}
\concept{Core}
\keyword{variable}
