% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_495.R
\name{wkathome}
\alias{wkathome}
\title{How often do you work from home}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkathome
}
\details{
Question (HAND CARD D8)
And how often...
...do you work at home during your usual working hours?
Would you say...
}
\section{Values}{

\itemize{
\item \code{1} always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} hardly ever
\item \code{5} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6330/vshow}{GSS Data Explorer page for \code{wkathome}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab always \tab often \tab sometimes \tab hardly ever \tab never \tab no answer \tab Total \cr
   2016 \tab 1927 \tab 64 \tab 83 \tab 142 \tab 162 \tab 484 \tab 5 \tab 2867 \cr
   Total \tab 1927 \tab 64 \tab 83 \tab 142 \tab 162 \tab 484 \tab 5 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{chngewrk}},
\code{\link{decidwrk}},
\code{\link{harass5}},
\code{\link{nounemp5}},
\code{\link{nounemp6}},
\code{\link{rcontact}},
\code{\link{unionsbd}},
\code{\link{wkwkends}},
\code{\link{wrkshift}}
}
\concept{Work Orientation}
\keyword{variable}
