% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_337.R
\name{wkbhvrs}
\alias{wkbhvrs}
\title{Employee doing these things at work to r}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkbhvrs
}
\details{
Question In the past 12 months, the person who has engaged in these types of behaviors has most often been:
}
\section{Values}{

\itemize{
\item \code{1} a co-worker
\item \code{2} your supervisor or boss
\item \code{3} a customer/client/patient
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3022/vshow}{GSS Data Explorer page for \code{wkbhvrs}}.
}

\seealso{
Other Workplace Conflict: 
\code{\link{denyrais}},
\code{\link{eharaswk}},
\code{\link{ignorwk}},
\code{\link{jokeswk}},
\code{\link{liedcwkr}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{wkrspns}}

Other Employment: 
\code{\link{denyrais}},
\code{\link{eharaswk}},
\code{\link{ignorwk}},
\code{\link{jokeswk}},
\code{\link{liedcwkr}},
\code{\link{rudewk}},
\code{\link{rumorwk}}

Other Job: 
\code{\link{denyrais}},
\code{\link{eharaswk}},
\code{\link{ignorwk}},
\code{\link{jokeswk}},
\code{\link{liedcwkr}},
\code{\link{rudewk}},
\code{\link{rumorwk}}

Other Work: 
\code{\link{denyrais}},
\code{\link{eharaswk}},
\code{\link{ignorwk}},
\code{\link{jokeswk}},
\code{\link{liedcwkr}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{wkrspns}}
}
\concept{Employment}
\concept{Job}
\concept{Work}
\concept{Workplace Conflict}
\keyword{variable}
