% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_309.R
\name{wkbonus}
\alias{wkbonus}
\title{R is likely to get a bonus or pay increase}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkbonus
}
\details{
Question When you do your job well, are you likely to get a bonus or pay increase?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} maybe
\item \code{3} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2815/vshow}{GSS Data Explorer page for \code{wkbonus}}.
}

\seealso{
Other Quality of Working Life: 
\code{\link{cowrkhlp}},
\code{\link{fairearn}},
\code{\link{handmove}},
\code{\link{hvylift}},
\code{\link{manvsemp}},
\code{\link{suphelp}},
\code{\link{trainops}},
\code{\link{wkpraise}},
\code{\link{wrktime}}
}
\concept{Quality of Working Life}
\keyword{variable}
