% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_424.R
\name{wkcomptr}
\alias{wkcomptr}
\title{Use computer at work}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkcomptr
}
\details{
Question Do you use a computer or computerized equipment at work?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3887/vshow}{GSS Data Explorer page for \code{wkcomptr}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   2008 \tab 1165 \tab 642 \tab 202 \tab 1 \tab 13 \tab 2023 \cr
   Total \tab 1165 \tab 642 \tab 202 \tab 1 \tab 13 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{findnwjb}},
\code{\link{losejb12}},
\code{\link{lostjobs}},
\code{\link{moretrde}},
\code{\link{newjobs}},
\code{\link{whyjbct2}},
\code{\link{whylose1}},
\code{\link{whylose2}},
\code{\link{whylose3}}
}
\concept{Global Economics}
\keyword{variable}
