% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_601.R
\name{wkcontct}
\alias{wkcontct}
\title{How often contacted about work when not working}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkcontct
}
\details{
Question Some workers must respond to coworkers, supervisors, managers, customers, or clients about work-related matters when they are not working. These contacts may be by phone, text, instant message, email, etc. How often are you contacted about work-related matters when you are not working?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} less than once a month
\item \code{3} once or twice a month
\item \code{4} once a week
\item \code{5} several times a week
\item \code{6} once a day
\item \code{7} two or more times a day
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2014 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5504/vshow}{GSS Data Explorer page for \code{wkcontct}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab never \tab less than once a month \tab once or twice a month \tab once a week \tab several times a week \tab once a day \tab two or more times a day \tab don't know \tab no answer \tab Total \cr
   2014 \tab 1292 \tab 401 \tab 214 \tab 194 \tab 119 \tab 202 \tab 42 \tab 65 \tab 7 \tab 2 \tab 2538 \cr
   Total \tab 1292 \tab 401 \tab 214 \tab 194 \tab 119 \tab 202 \tab 42 \tab 65 \tab 7 \tab 2 \tab 2538 \cr
}
}

\seealso{
Other Work and Relationships: 
\code{\link{effctsup}},
\code{\link{talkspvs}}
}
\concept{Work and Relationships}
\keyword{variable}
