% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_601.R
\name{wkcontct}
\alias{wkcontct}
\title{How often contacted about work when not working}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkcontct
}
\details{
Question Some workers must respond to coworkers, supervisors, managers, customers, or clients about work-related matters when they are not working. These contacts may be by phone, text, instant message, email, etc. How often are you contacted about work-related matters when you are not working?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} less than once a month
\item \code{3} once or twice a month
\item \code{4} once a week
\item \code{5} several times a week
\item \code{6} once a day
\item \code{7} two or more times a day
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2014 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/5504/vshow}{GSS Data Explorer page for \code{wkcontct}}.
}

\seealso{
Other Work and Relationships: 
\code{\link{effctsup}},
\code{\link{talkspvs}}
}
\concept{Work and Relationships}
\keyword{variable}
