% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_199.R
\name{wkfambiz}
\alias{wkfambiz}
\title{Work for pay in a farm etc owned by your family?}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkfambiz
}
\details{
Question Do you work for pay in a farm or other business owned in whole or in substantial part by a member or members of your family (parents, grandparents, children, brothers, sisters, aunts, uncles, or first cousins)?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1876/vshow}{GSS Data Explorer page for \code{wkfambiz}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1996 \tab 1460 \tab 54 \tab 1381 \tab 9 \tab 2904 \cr
   Total \tab 1460 \tab 54 \tab 1381 \tab 9 \tab 2904 \cr
}
}

\seealso{
Other Markets: 
\code{\link{decauto}},
\code{\link{econsys}},
\code{\link{fambiz}},
\code{\link{famfinan}},
\code{\link{haggle}},
\code{\link{hrfambiz}},
\code{\link{natrecon}},
\code{\link{reqinfo}},
\code{\link{tagsales}}
}
\concept{Markets}
\keyword{variable}
