% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_199.R
\name{wkfambiz}
\alias{wkfambiz}
\title{Work for pay in a farm etc owned by your family?}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkfambiz
}
\details{
Question Do you work for pay in a farm or other business owned in whole or in substantial part by a member or members of your family (parents, grandparents, children, brothers, sisters, aunts, uncles, or first cousins)?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1996 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/1876/vshow}{GSS Data Explorer page for \code{wkfambiz}}.
}

\seealso{
Other Markets: 
\code{\link{decauto}},
\code{\link{econsys}},
\code{\link{fambiz}},
\code{\link{famfinan}},
\code{\link{haggle}},
\code{\link{hrfambiz}},
\code{\link{natrecon}},
\code{\link{reqinfo}},
\code{\link{tagsales}}

Other Agriculture: 
\code{\link{fambiz}}

Other Labor Force Status: 
\code{\link{hrfambiz}}

Other Occupation: 
\code{\link{fambiz}}
}
\concept{Agriculture}
\concept{Labor Force Status}
\concept{Markets}
\concept{Occupation}
\concept{Work}
\keyword{variable}
