% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_490.R
\name{wkkidscs}
\alias{wkkidscs}
\title{Did partner work outside home with child under school age}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkkidscs
}
\details{
Question What about your (spouse/partner) at that time " did (he/she) work outside the home full-time, part-time, or not at all…
A. When a child was under school age?
}
\section{Values}{

\itemize{
\item \verb{[1]} worked full-time
\item \verb{[2]} worked part-time
\item \verb{[3]} stayed at home
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2012 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4481/vshow}{GSS Data Explorer page for \code{wkkidscs}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab worked full-time \tab worked part-time \tab stayed at home \tab don't know \tab no answer \tab Total \cr
   2012 \tab 1044 \tab 557 \tab 91 \tab 177 \tab 95 \tab 10 \tab 1974 \cr
   Total \tab 1044 \tab 557 \tab 91 \tab 177 \tab 95 \tab 10 \tab 1974 \cr
}
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{eldcost}},
\code{\link{famvswk1}},
\code{\link{hhclean1}},
\code{\link{jobvsfa1}},
\code{\link{tiredhm1}},
\code{\link{tiredwk1}},
\code{\link{wkkidscl}},
\code{\link{wkndact}},
\code{\link{wkyngscl}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
