% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_490.R
\name{wkndact}
\alias{wkndact}
\title{Who decides weekend activities}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkndact
}
\details{
Question When you and your (spouse/partner) make decisions about choosing shared weekend activities, who has the final say?
}
\section{Values}{

\itemize{
\item \code{1} mostly me
\item \code{2} mostly my spousepartfill
\item \code{3} sometimes me/sometimes my spousepartfill
\item \code{4} we decide together
\item \code{5} someone else
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/4474/vshow}{GSS Data Explorer page for \code{wkndact}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{eldcost}},
\code{\link{famvswk1}},
\code{\link{hhclean1}},
\code{\link{jobvsfa1}},
\code{\link{tiredhm1}},
\code{\link{tiredwk1}},
\code{\link{wkkidscl}},
\code{\link{wkkidscs}},
\code{\link{wkyngscl}}
}
\concept{ISSP Family & Gender Roles}
\concept{Marriage}
\keyword{variable}
