% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_309.R
\name{wkpraise}
\alias{wkpraise}
\title{R is likely to be praised by supervisor}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkpraise
}
\details{
Question When you do your job well, are you likely to be praised by your supervisor or employer?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} maybe
\item \code{3} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010, 2014, 2018, 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2814/vshow}{GSS Data Explorer page for \code{wkpraise}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes \tab maybe \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2002 \tab 969 \tab 954 \tab 453 \tab 348 \tab 17 \tab 24 \tab - \tab 2765 \cr
   2006 \tab 2776 \tab 926 \tab 451 \tab 291 \tab 40 \tab 26 \tab - \tab 4510 \cr
   2010 \tab 857 \tab 609 \tab 320 \tab 210 \tab 14 \tab 34 \tab - \tab 2044 \cr
   2014 \tab 1288 \tab 669 \tab 343 \tab 214 \tab 10 \tab 14 \tab - \tab 2538 \cr
   2018 \tab 900 \tab 765 \tab 418 \tab 215 \tab 14 \tab 36 \tab - \tab 2348 \cr
   2022 \tab 1555 \tab 1036 \tab 589 \tab 307 \tab 17 \tab 21 \tab 19 \tab 3544 \cr
   Total \tab 8345 \tab 4959 \tab 2574 \tab 1585 \tab 112 \tab 155 \tab 19 \tab 17749 \cr
}
}

\seealso{
Other Quality of Working Life: 
\code{\link{cowrkhlp}},
\code{\link{fairearn}},
\code{\link{handmove}},
\code{\link{hvylift}},
\code{\link{manvsemp}},
\code{\link{suphelp}},
\code{\link{trainops}},
\code{\link{wkbonus}},
\code{\link{wrktime}}
}
\concept{Quality of Working Life}
\keyword{variable}
