% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_337.R
\name{wkrspns}
\alias{wkrspns}
\title{Response to misbehavior from coworkers}
\source{
General Social Survey https://gss.norc.org
}
\description{
wkrspns
}
\details{
Question What did you typically do in response to these behaviors?
}
\section{Values}{

\itemize{
\item \code{1} did nothing (ignored it)
\item \code{2} talked to co-workers or other people at work
\item \code{3} talked to family or friends
\item \code{4} reported the incident to management or those higher up
\item \code{5} reported the incident to the union
\item \code{6} confronted the person (s) engaging in the behaviors
\item \code{7} filed a formal complaint or grievance about the behavior or experiences
\item \code{8} did not know what to do
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2012 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3023/vshow}{GSS Data Explorer page for \code{wkrspns}}.
}

\seealso{
Other Workplace Conflict: 
\code{\link{actupset}},
\code{\link{denyrais}},
\code{\link{difstand}},
\code{\link{eharaswk}},
\code{\link{gdjobsec}},
\code{\link{health1}},
\code{\link{hotargus}},
\code{\link{ignorwk}},
\code{\link{jobsecok}},
\code{\link{jokeswk}},
\code{\link{lackinfo}},
\code{\link{liedcwkr}},
\code{\link{lookaway}},
\code{\link{manvsemp}},
\code{\link{mntlhlth}},
\code{\link{numemps}},
\code{\link{numemps_0422}},
\code{\link{othcredt}},
\code{\link{perspace}},
\code{\link{physharm}},
\code{\link{putdown}},
\code{\link{reliedon}},
\code{\link{rptprobs}},
\code{\link{rudewk}},
\code{\link{rumorwk}},
\code{\link{shout}},
\code{\link{skipwork}},
\code{\link{supcares}},
\code{\link{treatres}},
\code{\link{trynewjb}},
\code{\link{wkageism}},
\code{\link{wkbhvrs}},
\code{\link{wkracism}},
\code{\link{wkvsfam}}
}
\concept{Workplace Conflict}
\keyword{variable}
