% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_310.R
\name{wksexism}
\alias{wksexism}
\title{R feels discriminated because of gender}
\source{
General Social Survey https://gss.norc.org
}
\description{
wksexism
}
\details{
Question Do you feel in any way discriminated against on your job because of your gender?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2823/vshow}{GSS Data Explorer page for \code{wksexism}}.
}

\seealso{
Other Quality of Working Life: 
\code{\link{health1}},
\code{\link{jobfind1}},
\code{\link{laidoff}},
\code{\link{rincblls}},
\code{\link{trynewjb}},
\code{\link{wkageism}},
\code{\link{wkharoth}},
\code{\link{wkharsex}},
\code{\link{wkracism}}
}
\concept{Quality of Working Life}
\keyword{variable}
