% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_64.R
\name{wksub1}
\alias{wksub1}
\title{Does r or spouse or partner have a supervisor}
\source{
General Social Survey https://gss.norc.org
}
\description{
wksub1
}
\details{
Question Do you (does your (SPOUSE or PARTNER) have a supervisor on (your/his/her) job to whom you are (he/she/is) directly responsible?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2016, 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6176/vshow}{GSS Data Explorer page for \code{wksub1}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab 6 \tab Total \cr
   2016 \tab 914 \tab 1588 \tab 359 \tab 5 \tab 1 \tab - \tab - \tab 2867 \cr
   2018 \tab 693 \tab 1356 \tab 293 \tab 5 \tab 1 \tab - \tab - \tab 2348 \cr
   2021 \tab 1419 \tab 2132 \tab 434 \tab 40 \tab 1 \tab 6 \tab - \tab 4032 \cr
   2022 \tab 1203 \tab 1910 \tab 398 \tab - \tab 6 \tab 5 \tab 22 \tab 3544 \cr
   Total \tab 4229 \tab 6986 \tab 1484 \tab 50 \tab 9 \tab 11 \tab 22 \tab 12791 \cr
}
}

\seealso{
Other Core: 
\code{\link{finalter}},
\code{\link{finrela}},
\code{\link{incneed}},
\code{\link{mininc}},
\code{\link{satfin}},
\code{\link{wksub}},
\code{\link{wksubs}},
\code{\link{wksubs1}},
\code{\link{wksup}}
}
\concept{Core}
\keyword{variable}
