% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_64.R
\name{wksup}
\alias{wksup}
\title{Does r or spouse supervise anyone}
\source{
General Social Survey https://gss.norc.org
}
\description{
wksup
}
\details{
Question In your (SPOUSE'S) job, (do you/does he/she) supervise anyone who is directly responsible to (you/him/her)?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/-/- \tab full \cr
   1973 \tab -/-/-/- \tab full \cr
   1974 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/C/-/- \tab partial \cr
   1989 \tab A/C/-/- \tab partial \cr
   1990 \tab A/C/-/- \tab partial \cr
   1991 \tab A/C/-/- \tab partial \cr
   1993 \tab A/C/-/- \tab partial \cr
   1994 \tab A/C/-/- \tab partial \cr
   1996 \tab A/C/-/- \tab partial \cr
   1998 \tab A/C/-/- \tab partial \cr
   2000 \tab A/C/-/- \tab partial \cr
   2002 \tab A/C/-/- \tab partial \cr
   2004 \tab A/C/-/- \tab partial \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
   2024 \tab A/B/C/- \tab full \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/579/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab no \tab no answer \tab yes \tab don't know \tab skipped on web \tab Total \cr
   1972 \tab 416 \tab 730 \tab 52 \tab 415 \tab - \tab - \tab 1613 \cr
   1973 \tab 369 \tab 628 \tab 66 \tab 435 \tab 6 \tab - \tab 1504 \cr
   1974 \tab 400 \tab 604 \tab 51 \tab 426 \tab 3 \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab 450 \tab 572 \tab 102 \tab 374 \tab 1 \tab - \tab 1499 \cr
   1977 \tab 382 \tab 660 \tab 62 \tab 418 \tab 8 \tab - \tab 1530 \cr
   1978 \tab 1532 \tab - \tab - \tab - \tab - \tab - \tab 1532 \cr
   1980 \tab 404 \tab 575 \tab 25 \tab 460 \tab 4 \tab - \tab 1468 \cr
   1982 \tab 589 \tab 701 \tab 74 \tab 492 \tab 4 \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab 390 \tab 603 \tab 42 \tab 432 \tab 6 \tab - \tab 1473 \cr
   1985 \tab 438 \tab 601 \tab 20 \tab 468 \tab 7 \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab 506 \tab 739 \tab 53 \tab 512 \tab 9 \tab - \tab 1819 \cr
   1988 \tab 792 \tab 353 \tab 106 \tab 229 \tab 1 \tab - \tab 1481 \cr
   1989 \tab 808 \tab 392 \tab 23 \tab 310 \tab 4 \tab - \tab 1537 \cr
   1990 \tab 723 \tab 375 \tab 12 \tab 262 \tab - \tab - \tab 1372 \cr
   1991 \tab 841 \tab 418 \tab 17 \tab 239 \tab 2 \tab - \tab 1517 \cr
   1993 \tab 839 \tab 422 \tab 11 \tab 329 \tab 5 \tab - \tab 1606 \cr
   1994 \tab 1532 \tab 842 \tab 29 \tab 586 \tab 3 \tab - \tab 2992 \cr
   1996 \tab 1420 \tab 847 \tab 46 \tab 585 \tab 6 \tab - \tab 2904 \cr
   1998 \tab 1435 \tab 780 \tab 45 \tab 564 \tab 8 \tab - \tab 2832 \cr
   2000 \tab 1467 \tab 755 \tab 82 \tab 504 \tab 9 \tab - \tab 2817 \cr
   2002 \tab 1441 \tab 800 \tab 3 \tab 519 \tab 2 \tab - \tab 2765 \cr
   2004 \tab 1397 \tab 858 \tab 22 \tab 535 \tab - \tab - \tab 2812 \cr
   2006 \tab 1289 \tab 2000 \tab 17 \tab 1195 \tab 9 \tab - \tab 4510 \cr
   2008 \tab 581 \tab 881 \tab 7 \tab 549 \tab 5 \tab - \tab 2023 \cr
   2010 \tab 700 \tab 839 \tab 16 \tab 485 \tab 4 \tab - \tab 2044 \cr
   2012 \tab 607 \tab 837 \tab 11 \tab 507 \tab 12 \tab - \tab 1974 \cr
   2014 \tab 788 \tab 1117 \tab 4 \tab 625 \tab 4 \tab - \tab 2538 \cr
   2016 \tab 941 \tab 1175 \tab 17 \tab 729 \tab 5 \tab - \tab 2867 \cr
   2018 \tab 726 \tab 1029 \tab 11 \tab 578 \tab 4 \tab - \tab 2348 \cr
   2021 \tab 1445 \tab 1664 \tab 34 \tab 828 \tab 47 \tab 14 \tab 4032 \cr
   2022 \tab 1256 \tab 1492 \tab 10 \tab 761 \tab 17 \tab 8 \tab 3544 \cr
   2024 \tab 1250 \tab 1406 \tab 10 \tab 605 \tab 30 \tab 8 \tab 3309 \cr
   Total \tab 32713 \tab 25695 \tab 1080 \tab 15956 \tab 225 \tab 30 \tab 75699 \cr
}
}

\seealso{
Other Core: 
\code{\link{finalter}},
\code{\link{finrela}},
\code{\link{incneed}},
\code{\link{mininc}},
\code{\link{satfin}},
\code{\link{wksub}},
\code{\link{wksub1}},
\code{\link{wksubs}},
\code{\link{wksubs1}}

Other Job: 
\code{\link{wksub}}

Other Work: 
\code{\link{wksub}},
\code{\link{wksubs}}
}
\concept{Core}
\concept{Job}
\concept{Work}
\keyword{variable}
