% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_65.R
\name{wksups}
\alias{wksups}
\title{Does subordinate supervise anyone}
\source{
General Social Survey https://gss.norc.org
}
\description{
wksups
}
\details{
Question In your (SPOUSE''S) job, (do you/does he/she) supervise anyone who is directly responsible to (you/him/her)? A. IF YES: Do any of those persons supervise anyone else?
}
\section{Values}{

\itemize{
\item \code{3} yes
\item \code{4} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1972 \tab -/-/-/- \tab full \cr
   1973 \tab -/-/-/- \tab full \cr
   1974 \tab -/-/-/- \tab full \cr
   1976 \tab -/-/-/- \tab full \cr
   1977 \tab -/-/-/- \tab full \cr
   1980 \tab -/-/-/- \tab full \cr
   1982 \tab -/-/-/- \tab full \cr
   1984 \tab -/-/-/- \tab full \cr
   1985 \tab -/-/-/- \tab full \cr
   1987 \tab -/-/-/- \tab full \cr
   1988 \tab A/C/-/- \tab partial \cr
   1989 \tab A/C/-/- \tab partial \cr
   1990 \tab A/C/-/- \tab partial \cr
   1991 \tab A/C/-/- \tab partial \cr
   1993 \tab A/C/-/- \tab partial \cr
   1994 \tab A/C/-/- \tab partial \cr
   1996 \tab A/C/-/- \tab partial \cr
   1998 \tab A/C/-/- \tab partial \cr
   2000 \tab A/C/-/- \tab partial \cr
   2002 \tab A/C/-/- \tab partial \cr
   2004 \tab A/C/-/- \tab partial \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
   2021 \tab A/B/C/- \tab full \cr
   2022 \tab A/B/C/- \tab full \cr
   2024 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/580/vshow}{GSS Data Explorer page for \code{wksups}}.
}

\seealso{
Other Core: 
\code{\link{getahead}},
\code{\link{getaid}},
\code{\link{govaid}},
\code{\link{parsol}},
\code{\link{unemp}},
\code{\link{union}()},
\code{\link{union1}},
\code{\link{wksup1}},
\code{\link{wksups1}}

Other Work: 
\code{\link{getahead}}
}
\concept{Core}
\concept{Work}
\keyword{variable}
