% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_504.R
\name{wktenure}
\alias{wktenure}
\title{Describe your present job situation}
\source{
General Social Survey https://gss.norc.org
}
\description{
wktenure
}
\details{
Question Which of the following describes your present job situation?
}
\section{Values}{

\itemize{
\item \code{1} no writtn cntrct
\item \code{2} fixed less 1 yr
\item \code{3} fixed more 1 yr
\item \code{4} job no tme limit
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4581/vshow}{GSS Data Explorer page for \code{wktenure}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab no writtn cntrct \tab fixed less 1 yr \tab fixed more 1 yr \tab job no tme limit \tab can't choose \tab no answer \tab Total \cr
   1998 \tab 1967 \tab 335 \tab 37 \tab 53 \tab 368 \tab 43 \tab 29 \tab 2832 \cr
   Total \tab 1967 \tab 335 \tab 37 \tab 53 \tab 368 \tab 43 \tab 29 \tab 2832 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{absent}},
\code{\link{chngwork}},
\code{\link{edcskill}},
\code{\link{helporg1}},
\code{\link{jobskill}},
\code{\link{prideorg}},
\code{\link{proudwrk}},
\code{\link{stayorg3}},
\code{\link{useskill}}
}
\concept{Work Orientation}
\keyword{variable}
