% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_175.R
\name{wlthimm}
\alias{wlthimm}
\title{Are legal immigrants rich or poor}
\source{
General Social Survey https://gss.norc.org
}
\description{
wlthimm
}
\details{
Question Now I have some questions about different groups in our society. I'm going to show you a seven-point scale which the characteristics of people in a group can be rated. In the first statement a score of 1 means that you think almost all of the people in that group are "rich."  A score of 7 means that you think almost everyone in the group are "poor."  A score of 4 means that you think that the group is not towards one end or another, and of course you may choose any number in between that comes closest to where you think people in the group stand.
A1. Immigrants who are in this country legally?
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 - rich
\item \verb{[7]} 7 - poor
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1636/vshow}{GSS Data Explorer page for \code{wlthimm}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab 1 - rich \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 - poor \tab don't know \tab no answer \tab Total \cr
   1994 \tab 1518 \tab 15 \tab 48 \tab 118 \tab 420 \tab 369 \tab 261 \tab 110 \tab 119 \tab 14 \tab 2992 \cr
   Total \tab 1518 \tab 15 \tab 48 \tab 118 \tab 420 \tab 369 \tab 261 \tab 110 \tab 119 \tab 14 \tab 2992 \cr
}
}

\seealso{
Other Multiculturalism: 
\code{\link{asngovt}},
\code{\link{blkgovt}},
\code{\link{econpast}},
\code{\link{hspgovt}},
\code{\link{obrespct}},
\code{\link{whtgovt}},
\code{\link{wlthundc}},
\code{\link{workimm}},
\code{\link{workundc}}
}
\concept{Multiculturalism}
\keyword{variable}
