% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_295.R
\name{wlthpov}
\alias{wlthpov}
\title{Some rich some poor ok in free society}
\source{
General Social Survey https://gss.norc.org
}
\description{
wlthpov
}
\details{
Question Please tell me whether you strongly agree, agree, neither agree nor disagree, or strongly disagree with the following statement: In a free society, it is all right if a few people accumulate a lot of wealth and property while many others live  in poverty.
}
\section{Values}{

\itemize{
\item \verb{[1]} strongly agree
\item \verb{[2]} agree
\item \verb{[3]} neither agree nor disagree
\item \verb{[4]} disagree
\item \verb{[5]} strongly disagree
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2675/vshow}{GSS Data Explorer page for \code{wlthpov}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab neither agree nor disagree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab Total \cr
   2000 \tab 1398 \tab 114 \tab 410 \tab 264 \tab 370 \tab 208 \tab 37 \tab 16 \tab 2817 \cr
   Total \tab 1398 \tab 114 \tab 410 \tab 264 \tab 370 \tab 208 \tab 37 \tab 16 \tab 2817 \cr
}
}

\seealso{
Other Freedom: 
\code{\link{cntrlife}},
\code{\link{expunpop}},
\code{\link{freeprss}}
}
\concept{Freedom}
\keyword{variable}
