% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_126.R
\name{wlthwhts}
\alias{wlthwhts}
\title{Rich - poor}
\source{
General Social Survey https://gss.norc.org
}
\description{
wlthwhts
}
\details{
Question Now I have some questions about different groups in our society. I'm going to show you a seven-point scale on which the characteristics of people in a group can be rated. In the first statement a score of 1 means that you think almost all of the people in that group are "rich."  A score of 7 means that you think almost everyone in the group are "poor."  A score of 4 means you think that the group is not towards one end or another, and of course you may choose any number In between that comes closest to where you think people in the group stand.
\enumerate{
\item Where would you rate whites in general on this scale?
}
}
\section{Values}{

\itemize{
\item \verb{[1]} 1 - rich
\item \verb{[7]} 7 - poor
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1996, 1998, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018, 2021, 2022, 2024 \cr
   A/B/C \tab 1990, 1994, 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1178/vshow}{GSS Data Explorer page for \code{wlthwhts}}.

Counts by year:\tabular{lllllllllllll}{
   year \tab iap \tab 1 - rich \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab 7 - poor \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   1990 \tab - \tab 56 \tab 84 \tab 398 \tab 693 \tab 77 \tab 5 \tab 3 \tab 40 \tab 16 \tab - \tab 1372 \cr
   1994 \tab 1518 \tab 56 \tab 118 \tab 375 \tab 765 \tab 84 \tab 19 \tab 6 \tab 38 \tab 13 \tab - \tab 2992 \cr
   1996 \tab 1935 \tab 41 \tab 61 \tab 248 \tab 485 \tab 82 \tab 10 \tab 2 \tab 34 \tab 6 \tab - \tab 2904 \cr
   1998 \tab 961 \tab 71 \tab 159 \tab 534 \tab 878 \tab 121 \tab 25 \tab 7 \tab 66 \tab 10 \tab - \tab 2832 \cr
   2000 \tab 486 \tab 122 \tab 255 \tab 670 \tab 976 \tab 175 \tab 35 \tab 8 \tab 69 \tab 21 \tab - \tab 2817 \cr
   2002 \tab 1857 \tab 36 \tab 93 \tab 279 \tab 393 \tab 73 \tab 7 \tab 6 \tab 18 \tab 3 \tab - \tab 2765 \cr
   2004 \tab 1906 \tab 43 \tab 83 \tab 247 \tab 449 \tab 59 \tab 11 \tab - \tab 12 \tab 2 \tab - \tab 2812 \cr
   2006 \tab 2518 \tab 105 \tab 195 \tab 573 \tab 888 \tab 150 \tab 22 \tab 10 \tab 40 \tab 9 \tab - \tab 4510 \cr
   2008 \tab 694 \tab 71 \tab 87 \tab 353 \tab 662 \tab 105 \tab 11 \tab 9 \tab 25 \tab 6 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 73 \tab 123 \tab 361 \tab 700 \tab 115 \tab 13 \tab 5 \tab 30 \tab 10 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 65 \tab 90 \tab 332 \tab 681 \tab 91 \tab 12 \tab 2 \tab 20 \tab 9 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 89 \tab 157 \tab 412 \tab 857 \tab 107 \tab 17 \tab 10 \tab 17 \tab 9 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 95 \tab 186 \tab 458 \tab 985 \tab 110 \tab 22 \tab 7 \tab 16 \tab 9 \tab - \tab 2867 \cr
   2018 \tab 1170 \tab 68 \tab 112 \tab 305 \tab 585 \tab 73 \tab 12 \tab 5 \tab 11 \tab 7 \tab - \tab 2348 \cr
   2021 \tab 2689 \tab 70 \tab 133 \tab 388 \tab 653 \tab 68 \tab 19 \tab 5 \tab 2 \tab 1 \tab 4 \tab 4032 \cr
   2022 \tab 2339 \tab 95 \tab 142 \tab 326 \tab 526 \tab 67 \tab 11 \tab 11 \tab 19 \tab 7 \tab 1 \tab 3544 \cr
   Total \tab 21201 \tab 1156 \tab 2078 \tab 6259 \tab 11176 \tab 1557 \tab 251 \tab 96 \tab 457 \tab 138 \tab 5 \tab 44374 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{socdif1}},
\code{\link{socdif2}},
\code{\link{socdif3}},
\code{\link{socdif4}},
\code{\link{whypoor1}},
\code{\link{whypoor2}},
\code{\link{whypoor3}},
\code{\link{whypoor4}},
\code{\link{wlthjews}}
}
\concept{Intergroup Relations}
\keyword{variable}
