% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_425.R
\name{wocomptr}
\alias{wocomptr}
\title{R can do job without computer}
\source{
General Social Survey https://gss.norc.org
}
\description{
wocomptr
}
\details{
Question Could you do your job without a computer or computerized equipment? Would you say….
}
\section{Values}{

\itemize{
\item \code{1} couldn't do my job without a computer or computerized equipment
\item \code{2} could do my job without a computer or computerized equipment but it would be much harder
\item \code{3} doing my job without a computer or computerized equipment would be about the same as now
\item \code{4} already do most of my job without a computer or computerized equipment
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3888/vshow}{GSS Data Explorer page for \code{wocomptr}}.
}

\seealso{
Other Global Economics: 
\code{\link{autonojb}},
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{whynojb1}},
\code{\link{whynojb2}}

Other Computers And Internet: 
\code{\link{autonojb}}

Other Job: 
\code{\link{autonojb}},
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{whynojb1}},
\code{\link{whynojb2}}
}
\concept{Computers And Internet}
\concept{Global Economics}
\concept{Job}
\keyword{variable}
