% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_425.R
\name{wocomptr}
\alias{wocomptr}
\title{R can do job without computer}
\source{
General Social Survey https://gss.norc.org
}
\description{
wocomptr
}
\details{
Question Could you do your job without a computer or computerized equipment? Would you say….
}
\section{Values}{

\itemize{
\item \verb{[1]} couldn't do my job without a computer or computerized equipment
\item \verb{[2]} could do my job without a computer or computerized equipment but it would be much harder
\item \verb{[3]} doing my job without a computer or computerized equipment would be about the same as now
\item \verb{[4]} already do most of my job without a computer or computerized equipment
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2008 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3888/vshow}{GSS Data Explorer page for \code{wocomptr}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab couldn't do my job without a computer or computerized equipment \tab could do my job without a computer or computerized equipment but it would be much harder \tab doing my job without a computer or computerized equipment would be about the same as now \tab already do most of my job without a computer or computerized equipment \tab Total \cr
   2008 \tab 1381 \tab 289 \tab 248 \tab 52 \tab 53 \tab 2023 \cr
   Total \tab 1381 \tab 289 \tab 248 \tab 52 \tab 53 \tab 2023 \cr
}
}

\seealso{
Other Global Economics: 
\code{\link{autonojb}},
\code{\link{inclstjb}},
\code{\link{intlcowk}},
\code{\link{lastyrwk}},
\code{\link{leavejb}},
\code{\link{meetf2f1}},
\code{\link{meetf2f2}},
\code{\link{whynojb1}},
\code{\link{whynojb2}}
}
\concept{Global Economics}
\keyword{variable}
