% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_240.R
\name{wojob}
\alias{wojob}
\title{Since 1st job ever been without a job}
\source{
General Social Survey https://gss.norc.org
}
\description{
wojob
}
\details{
Question Since your first full-time job, has there ever been a time when you've been without a job and not looking for a job?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2157/vshow}{GSS Data Explorer page for \code{wojob}}.
}

\seealso{
Other Jobs: 
\code{\link{doingdys}},
\code{\link{doingfin}},
\code{\link{doingtrn}},
\code{\link{formldys}},
\code{\link{formlfin}},
\code{\link{formlhrs}},
\code{\link{howpaid}},
\code{\link{numemp}},
\code{\link{wojobyrs}}
}
\concept{Jobs}
\keyword{variable}
