% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_240.R
\name{wojob}
\alias{wojob}
\title{Since 1st job ever been without a job}
\source{
General Social Survey https://gss.norc.org
}
\description{
wojob
}
\details{
Question Since your first full-time job, has there ever been a time when you've been without a job and not looking for a job?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2157/vshow}{GSS Data Explorer page for \code{wojob}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1998 \tab 1455 \tab 495 \tab 854 \tab 28 \tab 2832 \cr
   Total \tab 1455 \tab 495 \tab 854 \tab 28 \tab 2832 \cr
}
}

\seealso{
Other Jobs: 
\code{\link{doingdys}},
\code{\link{doingfin}},
\code{\link{doingtrn}},
\code{\link{formldys}},
\code{\link{formlfin}},
\code{\link{formlhrs}},
\code{\link{howpaid}},
\code{\link{numemp}},
\code{\link{wojobyrs}}
}
\concept{Jobs}
\keyword{variable}
