% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_240.R
\name{wojobyrs}
\alias{wojobyrs}
\title{How many years without a job}
\source{
General Social Survey https://gss.norc.org
}
\description{
wojobyrs
}
\details{
Question Since your first full-time job, has there ever been a time when you''ve been without a job and not looking for a job?
B. How many years in total?
}
\section{Values}{

\itemize{
\item \code{1} less one year
\item \code{2} one to 3 years
\item \code{3} 3 to 5 years
\item \code{4} five - ten years
\item \code{5} ten years/over
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2158/vshow}{GSS Data Explorer page for \code{wojobyrs}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab less one year \tab one to 3 years \tab 3 to 5 years \tab five - ten years \tab ten years/over \tab don't know \tab no answer \tab Total \cr
   1998 \tab 2309 \tab 243 \tab 122 \tab 35 \tab 46 \tab 39 \tab 1 \tab 37 \tab 2832 \cr
   Total \tab 2309 \tab 243 \tab 122 \tab 35 \tab 46 \tab 39 \tab 1 \tab 37 \tab 2832 \cr
}
}

\seealso{
Other Jobs: 
\code{\link{doingdys}},
\code{\link{doingfin}},
\code{\link{doingtrn}},
\code{\link{formldys}},
\code{\link{formlfin}},
\code{\link{formlhrs}},
\code{\link{howpaid}},
\code{\link{numemp}},
\code{\link{wojob}}
}
\concept{Jobs}
\keyword{variable}
