% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_578.R
\name{wordh}
\alias{wordh}
\title{Word h}
\source{
General Social Survey https://gss.norc.org
}
\description{
wordh
}
\details{
Question We would like to know something about how people go about guessing words they do not know. On this card are listed some words--you may know some of them, and you may not know quite a few of them. On each line the first word is in capital letters -- like BEAST. Then there are five other words. Tell me the number of the word that comes closest to the meaning of the word in capital letters. For example, if the word in capital letters is BEAST, you would say "4" since "animal" come closer to BEAST than any of the other words. If you wish, I will read the words to you. These words are difficult for almost everyone -- just give me your best guess if you are not sure of the answer. CIRCLE ONE CODE NUMBER FOR EACH ITEM BELOW.
H. Word H
}
\section{Values}{

\itemize{
\item \code{0} incorrect answer
\item \code{1} correct answer
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5302/vshow}{GSS Data Explorer page for this variable}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab correct answer \tab incorrect answer \tab no answer \tab not available in this year \tab Total \cr
   1972 \tab 1613 \tab - \tab - \tab - \tab - \tab 1613 \cr
   1973 \tab 1504 \tab - \tab - \tab - \tab - \tab 1504 \cr
   1974 \tab - \tab 488 \tab 730 \tab 266 \tab - \tab 1484 \cr
   1975 \tab 1490 \tab - \tab - \tab - \tab - \tab 1490 \cr
   1976 \tab - \tab 485 \tab 692 \tab 322 \tab - \tab 1499 \cr
   1977 \tab 1530 \tab - \tab - \tab - \tab - \tab 1530 \cr
   1978 \tab - \tab 465 \tab 724 \tab 343 \tab - \tab 1532 \cr
   1980 \tab 1468 \tab - \tab - \tab - \tab - \tab 1468 \cr
   1982 \tab - \tab 452 \tab 908 \tab 500 \tab - \tab 1860 \cr
   1983 \tab 1599 \tab - \tab - \tab - \tab - \tab 1599 \cr
   1984 \tab - \tab 422 \tab 743 \tab 308 \tab - \tab 1473 \cr
   1985 \tab 1534 \tab - \tab - \tab - \tab - \tab 1534 \cr
   1986 \tab 1470 \tab - \tab - \tab - \tab - \tab 1470 \cr
   1987 \tab - \tab 441 \tab 844 \tab 534 \tab - \tab 1819 \cr
   1988 \tab 493 \tab 244 \tab 461 \tab 283 \tab - \tab 1481 \cr
   1989 \tab 531 \tab 270 \tab 459 \tab 277 \tab - \tab 1537 \cr
   1990 \tab 444 \tab 269 \tab 328 \tab 331 \tab - \tab 1372 \cr
   1991 \tab 493 \tab 273 \tab 443 \tab 308 \tab - \tab 1517 \cr
   1993 \tab 526 \tab 296 \tab 528 \tab 256 \tab - \tab 1606 \cr
   1994 \tab 1015 \tab 550 \tab 912 \tab 515 \tab - \tab 2992 \cr
   1996 \tab 944 \tab 505 \tab 876 \tab 579 \tab - \tab 2904 \cr
   1998 \tab 1445 \tab 356 \tab 577 \tab 454 \tab - \tab 2832 \cr
   2000 \tab 1398 \tab 358 \tab 555 \tab 506 \tab - \tab 2817 \cr
   2002 \tab 2765 \tab - \tab - \tab - \tab - \tab 2765 \cr
   2004 \tab 1340 \tab 472 \tab 788 \tab 212 \tab - \tab 2812 \cr
   2006 \tab 3041 \tab 473 \tab 675 \tab 321 \tab - \tab 4510 \cr
   2008 \tab 845 \tab 332 \tab 602 \tab 244 \tab - \tab 2023 \cr
   2010 \tab 614 \tab 449 \tab 722 \tab 259 \tab - \tab 2044 \cr
   2012 \tab 672 \tab 360 \tab 701 \tab 241 \tab - \tab 1974 \cr
   2014 \tab 863 \tab 513 \tab 932 \tab 230 \tab - \tab 2538 \cr
   2016 \tab 979 \tab 588 \tab 1089 \tab 211 \tab - \tab 2867 \cr
   2018 \tab 789 \tab 504 \tab 885 \tab 170 \tab - \tab 2348 \cr
   2021 \tab - \tab - \tab - \tab - \tab 4032 \tab 4032 \cr
   2022 \tab 1179 \tab 862 \tab 1179 \tab 324 \tab - \tab 3544 \cr
   2024 \tab 1126 \tab 799 \tab 1085 \tab 299 \tab - \tab 3309 \cr
   Total \tab 33710 \tab 11226 \tab 18438 \tab 8293 \tab 4032 \tab 75699 \cr
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1974 \tab -/-/- \tab full \cr
   1976 \tab -/-/- \tab full \cr
   1978 \tab -/-/- \tab full \cr
   1982 \tab -/-/- \tab full \cr
   1984 \tab -/-/- \tab full \cr
   1987 \tab -/-/- \tab full \cr
   1988 \tab A/B/- \tab partial \cr
   1989 \tab A/B/- \tab partial \cr
   1990 \tab A/B/- \tab partial \cr
   1991 \tab A/B/- \tab partial \cr
   1993 \tab A/B/- \tab partial \cr
   1994 \tab A/B/- \tab partial \cr
   1996 \tab A/B/- \tab partial \cr
   1998 \tab A/B/C \tab full \cr
   2000 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2008 \tab A/B/- \tab partial \cr
   2010 \tab A/B/- \tab partial \cr
   2012 \tab A/B/- \tab partial \cr
   2014 \tab A/B/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
   2022 \tab A/B/- \tab partial \cr
   2024 \tab A/B/- \tab partial \cr
}
}

\seealso{
Other Vocabulary: 
\code{\link{worda}},
\code{\link{wordb}},
\code{\link{wordc}},
\code{\link{wordd}},
\code{\link{worde}},
\code{\link{wordf}},
\code{\link{wordg}}

Other Knowledge: 
\code{\link{worda}},
\code{\link{wordb}},
\code{\link{wordc}},
\code{\link{wordd}},
\code{\link{worde}},
\code{\link{wordf}},
\code{\link{wordg}}

Other Vocabulary Test: 
\code{\link{worda}},
\code{\link{wordb}},
\code{\link{wordc}},
\code{\link{wordd}},
\code{\link{worde}},
\code{\link{wordf}},
\code{\link{wordg}}
}
\concept{Knowledge}
\concept{Vocabulary}
\concept{Vocabulary Test}
\keyword{variable}
