% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_497.R
\name{worecsup}
\alias{worecsup}
\title{Do you worry about losing economic support}
\source{
General Social Survey https://gss.norc.org
}
\description{
worecsup
}
\details{
Question To what extent, if at all, do you worry about the possibility of losing this main economic support?             
(Would you say...)
}
\section{Values}{

\itemize{
\item \verb{[1]} i worry a great deal
\item \verb{[2]} i worry to some extent
\item \verb{[3]} i worry a little
\item \verb{[4]} i don't worry at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6350/vshow}{GSS Data Explorer page for \code{worecsup}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab i worry a great deal \tab i worry to some extent \tab i worry a little \tab i don't worry at all \tab no answer \tab Total \cr
   2016 \tab 2326 \tab 104 \tab 88 \tab 122 \tab 203 \tab 24 \tab 2867 \cr
   Total \tab 2326 \tab 104 \tab 88 \tab 122 \tab 203 \tab 24 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{econsup1}},
\code{\link{moveaway}},
\code{\link{moveinus}},
\code{\link{unfamten}},
\code{\link{unjobexp}},
\code{\link{unmoney}},
\code{\link{unpeople}},
\code{\link{unrespct}},
\code{\link{unslfcon}}
}
\concept{Work Orientation}
\keyword{variable}
