% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_423.R
\name{work3yrs}
\alias{work3yrs}
\title{How many full time jobs r has in past 3 yrs since 2005}
\source{
General Social Survey https://gss.norc.org
}
\description{
work3yrs
}
\details{
Question How many full time jobs have you have had in past three years, that is since 2005 (counting the current job)?
}
\section{Values}{

\itemize{
\item \code{1} one
\item \code{2} number greater than 1
\item \code{3} none
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2008 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3872/vshow}{GSS Data Explorer page for \code{work3yrs}}.
}

\seealso{
Other Global Economics: 
\code{\link{curempyr}},
\code{\link{pastpay}},
\code{\link{paychnge}},
\code{\link{whyjbct1}},
\code{\link{whyleave}}

Other Employment: 
\code{\link{curempyr}},
\code{\link{whyjbct1}},
\code{\link{whyleave}}

Other Job: 
\code{\link{curempyr}},
\code{\link{pastpay}},
\code{\link{paychnge}},
\code{\link{whyjbct1}},
\code{\link{whyleave}}
}
\concept{Employment}
\concept{Global Economics}
\concept{Job}
\keyword{variable}
