% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_131.R
\name{workfare}
\alias{workfare}
\title{Work to get welfare}
\source{
General Social Survey https://gss.norc.org
}
\description{
workfare
}
\details{
Question Here are some proposals about welfare. I would like to know if you favor or oppose them. (READ "A") - Would you say that you strongly favor it, favor it, neither favor nor oppose it, oppose it, or strongly oppose it?
A. First, requiring that people must work in order to receive welfare.
}
\section{Values}{

\itemize{
\item \code{1} strongly favor
\item \code{2} favor
\item \code{3} neither favor nor oppose
\item \code{4} oppose
\item \code{5} strongly oppose
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1990 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1228/vshow}{GSS Data Explorer page for \code{workfare}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly favor \tab favor \tab neither favor nor oppose \tab oppose \tab strongly oppose \tab don't know \tab no answer \tab Total \cr
   1990 \tab - \tab 623 \tab 488 \tab 115 \tab 97 \tab 28 \tab 12 \tab 9 \tab 1372 \cr
   Total \tab 0 \tab 623 \tab 488 \tab 115 \tab 97 \tab 28 \tab 12 \tab 9 \tab 1372 \cr
}
}

\seealso{
Other Intergroup Relations: 
\code{\link{lessfare}},
\code{\link{livewhts}},
\code{\link{marasian}},
\code{\link{marblk}},
\code{\link{marhisp}},
\code{\link{marjew}},
\code{\link{marno}},
\code{\link{marso}},
\code{\link{marwht}}
}
\concept{Intergroup Relations}
\keyword{variable}
