% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_320.R
\name{workfor}
\alias{workfor}
\title{R work for whom}
\source{
General Social Survey https://gss.norc.org
}
\description{
workfor
}
\details{
Question Do you work for a private company, a non-profit organization, or for the government or a government agency?
}
\section{Values}{

\itemize{
\item \verb{[1]} private company
\item \verb{[2]} non-profit organization
\item \verb{[3]} government or government agency
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2002, 2006, 2010 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2850/vshow}{GSS Data Explorer page for \code{workfor}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab private company \tab non-profit organization \tab government or government agency \tab don't know \tab no answer \tab Total \cr
   2002 \tab 969 \tab 1370 \tab 114 \tab 305 \tab 1 \tab 6 \tab 2765 \cr
   2006 \tab 2778 \tab 1303 \tab 100 \tab 320 \tab 7 \tab 2 \tab 4510 \cr
   2010 \tab 857 \tab 834 \tab 105 \tab 201 \tab 18 \tab 29 \tab 2044 \cr
   Total \tab 4604 \tab 3507 \tab 319 \tab 826 \tab 26 \tab 37 \tab 9319 \cr
}
}

\seealso{
Other Employee Compensation: 
\code{\link{ownstock}},
\code{\link{stockops}},
\code{\link{stockval}},
\code{\link{workfor1}}
}
\concept{Employee Compensation}
\keyword{variable}
