% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_577.R
\name{workhr}
\alias{workhr}
\title{Hours r usually works}
\source{
General Social Survey https://gss.norc.org
}
\description{
workhr
}
\details{
Question What hours do you usually work--days, evenings, or nights?  CIRCLE ONE CODE.
}
\section{Values}{

\itemize{
\item \verb{[1]} reg day
\item \verb{[2]} eve
\item \verb{[3]} nights
\item \verb{[4]} varies
\item \verb{[5]} other
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1972 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/5288/vshow}{GSS Data Explorer page for \code{workhr}}.

Counts by year:\tabular{lllllllll}{
   year \tab reg day \tab eve \tab nights \tab varies \tab other \tab iap \tab no answer \tab Total \cr
   1972 \tab 557 \tab 24 \tab 23 \tab 87 \tab 54 \tab 863 \tab 5 \tab 1613 \cr
   Total \tab 557 \tab 24 \tab 23 \tab 87 \tab 54 \tab 863 \tab 5 \tab 1613 \cr
}
}

\seealso{
Other Military Service & Work: 
\code{\link{spdays}},
\code{\link{sphour}},
\code{\link{vetkind}},
\code{\link{vetyears}},
\code{\link{workdy}}
}
\concept{Military Service & Work}
\keyword{variable}
