% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_178.R
\name{worklife}
\alias{worklife}
\title{How successful in work life}
\source{
General Social Survey https://gss.norc.org
}
\description{
worklife
}
\details{
Question How successful do you feel in your work life? Do you feel...
}
\section{Values}{

\itemize{
\item \verb{[1]} not at all successful
\item \verb{[2]} not very successful
\item \verb{[3]} somewhat successful
\item \verb{[4]} very successful
\item \verb{[5]} completely successful
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1665/vshow}{GSS Data Explorer page for \code{worklife}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab not at all successful \tab not very successful \tab somewhat successful \tab very successful \tab completely successful \tab don't know \tab no answer \tab Total \cr
   1996 \tab 1910 \tab 15 \tab 73 \tab 443 \tab 366 \tab 89 \tab 3 \tab 5 \tab 2904 \cr
   Total \tab 1910 \tab 15 \tab 73 \tab 443 \tab 366 \tab 89 \tab 3 \tab 5 \tab 2904 \cr
}
}

\seealso{
Other Gender: 
\code{\link{balwkfam}},
\code{\link{earnmore}},
\code{\link{fairhwrk}},
\code{\link{famlife}},
\code{\link{noathome}},
\code{\link{refmorwk}},
\code{\link{refpromo}},
\code{\link{refxhour}},
\code{\link{workless}}
}
\concept{Gender}
\keyword{variable}
