% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_503.R
\name{worknow}
\alias{worknow}
\title{Are you currently working for pay}
\source{
General Social Survey https://gss.norc.org
}
\description{
worknow
}
\details{
Question Are you currently working for pay?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4610/vshow}{GSS Data Explorer page for \code{worknow}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   1998 \tab 1609 \tab 851 \tab 358 \tab 14 \tab 2832 \cr
   2006 \tab 2994 \tab 1016 \tab 500 \tab - \tab 4510 \cr
   2016 \tab 1390 \tab 936 \tab 537 \tab 4 \tab 2867 \cr
   Total \tab 5993 \tab 2803 \tab 1395 \tab 18 \tab 10209 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{paydojob}},
\code{\link{payeduc}},
\code{\link{payfam}},
\code{\link{paytime}},
\code{\link{placewrk}},
\code{\link{sethours}},
\code{\link{techjobs}},
\code{\link{techwork}},
\code{\link{wantjob1}}
}
\concept{Work Orientation}
\keyword{variable}
