% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_111.R
\name{workpol}
\alias{workpol}
\title{Has r worked for parties or candidates}
\source{
General Social Survey https://gss.norc.org
}
\description{
workpol
}
\details{
Question Have you done (other) work for one of the parties or candidates in most elections, some elections, only a few, or have you never done such work?
}
\section{Values}{

\itemize{
\item \code{1} most elections
\item \code{2} some elections
\item \code{3} only a few
\item \code{4} never
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1987 \tab - \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1029/vshow}{GSS Data Explorer page for \code{workpol}}.
}

\seealso{
Other Socio-political participation: 
\code{\link{intpol}},
\code{\link{lobbygrp}},
\code{\link{loccare}},
\code{\link{locgrp}},
\code{\link{locinflu}},
\code{\link{loclived}},
\code{\link{locprob}},
\code{\link{loctrust}},
\code{\link{swayvote}}
}
\concept{Socio-political participation}
\keyword{variable}
