% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_127.R
\name{workwhts}
\alias{workwhts}
\title{Hard working - lazy}
\source{
General Social Survey https://gss.norc.org
}
\description{
workwhts
}
\details{
Question Now I have some questions about different groups in our society. I'm going to show you a seven-point scale on which the characteristics of people in a group can be rated. In the first statement a score of 1 means that you think almost all of the people in that group are "rich."  A score of 7 means that you think almost everyone in the group are "poor."  A score of 4 means you think that the group is not towards one end or another, and of course you may choose any number In between that comes closest to where you think people in the group stand.
B. The second set of characteristics asks if people in the group tend to be hard-working or if they tend to be lazy.
\enumerate{
\item Where would you rate whites in general on this scale?
}
}
\section{Values}{

\itemize{
\item \code{1} 1 - hard-working
\item \code{7} 7 - lazy
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1990 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   1996 \tab A/B/- \tab partial \cr
   1998 \tab A/B/- \tab partial \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/- \tab partial \cr
   2004 \tab A/B/- \tab partial \cr
   2006 \tab A/B/- \tab partial \cr
   2008 \tab A/B/- \tab partial \cr
   2010 \tab A/B/- \tab partial \cr
   2012 \tab A/B/- \tab partial \cr
   2014 \tab A/B/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
   2021 \tab A/B/- \tab partial \cr
   2022 \tab A/B/- \tab partial \cr
   2024 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1184/vshow}{GSS Data Explorer page for \code{workwhts}}.
}

\seealso{
Other Intergroup Relations: 
\code{\link{wlthasns}},
\code{\link{wlthblks}},
\code{\link{wlthhsps}},
\code{\link{wlthso}},
\code{\link{workasns}},
\code{\link{workblks}},
\code{\link{workhsps}},
\code{\link{workjews}},
\code{\link{workso}}
}
\concept{Intergroup Relations}
\keyword{variable}
