% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_43.R
\name{world1}
\alias{world1}
\title{World image:much evil vs. much good}
\source{
General Social Survey https://gss.norc.org
}
\description{
world1
}
\details{
Question People have different images of the world and human nature. We'd like to know the kinds of images you have. Here is a card with sets of contrasting images. On a scale of 1-7 where would you place your image of the world  and human nature between the two contrasting images? (HAND CARD AH)
Look at the first set of contrasting images. If you think that "The world is basically filled with evil and sin," you would place yourself at 1. If you think "There is much goodness in the world which hints at God's goodness" you would place yourself at 7. If you think things are somewhere between these two you would place yourself at 2, 3,4, 5 or 6. (REPEAT EXAMPLE AS NECESSARY FOR EACH ITEM A-F) Where would you place your image of the world on the scale for . . .?
(READ EACH SET AND CIRCLE ONE CODE FOR EACH)
}
\section{Values}{

\itemize{
\item \code{1} world is evil
\item \code{7} world is good
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/- \tab 1989, 1990, 1994, 1996, 2000, 2002 \cr
   A/B/C \tab 1988, 1991, 1998 \cr
   Full \tab 1985, 1987 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/366/vshow}{GSS Data Explorer page for \code{world1}}.

Counts by year:\tabular{llllllllllll}{
   year \tab iap \tab world is evil \tab 2 \tab 3 \tab 4 \tab 5 \tab 6 \tab world is good \tab don't know \tab no answer \tab Total \cr
   1985 \tab - \tab 100 \tab 41 \tab 105 \tab 447 \tab 377 \tab 194 \tab 242 \tab 7 \tab 21 \tab 1534 \cr
   1987 \tab - \tab 143 \tab 68 \tab 154 \tab 461 \tab 372 \tab 225 \tab 365 \tab - \tab 31 \tab 1819 \cr
   1988 \tab - \tab 103 \tab 54 \tab 112 \tab 428 \tab 340 \tab 180 \tab 237 \tab 7 \tab 20 \tab 1481 \cr
   1989 \tab 531 \tab 66 \tab 38 \tab 77 \tab 290 \tab 229 \tab 134 \tab 151 \tab - \tab 21 \tab 1537 \cr
   1990 \tab 444 \tab 47 \tab 28 \tab 76 \tab 279 \tab 243 \tab 126 \tab 116 \tab 2 \tab 11 \tab 1372 \cr
   1991 \tab 158 \tab 74 \tab 52 \tab 73 \tab 339 \tab 326 \tab 237 \tab 169 \tab 3 \tab 86 \tab 1517 \cr
   1994 \tab 1522 \tab 78 \tab 61 \tab 155 \tab 418 \tab 345 \tab 196 \tab 196 \tab 1 \tab 20 \tab 2992 \cr
   1996 \tab 944 \tab 107 \tab 65 \tab 179 \tab 553 \tab 467 \tab 298 \tab 268 \tab 5 \tab 18 \tab 2904 \cr
   1998 \tab 463 \tab 125 \tab 92 \tab 196 \tab 687 \tab 542 \tab 384 \tab 307 \tab 7 \tab 29 \tab 2832 \cr
   2000 \tab 940 \tab 71 \tab 74 \tab 135 \tab 503 \tab 465 \tab 327 \tab 278 \tab 4 \tab 20 \tab 2817 \cr
   2002 \tab 1857 \tab 38 \tab 26 \tab 75 \tab 253 \tab 212 \tab 139 \tab 148 \tab 14 \tab 3 \tab 2765 \cr
   Total \tab 6859 \tab 952 \tab 599 \tab 1337 \tab 4658 \tab 3918 \tab 2440 \tab 2477 \tab 50 \tab 280 \tab 23570 \cr
}
}

\seealso{
Other Core: 
\code{\link{contemp}},
\code{\link{libtemp}},
\code{\link{prottemp}},
\code{\link{world2}},
\code{\link{world3}},
\code{\link{world4}},
\code{\link{world5}},
\code{\link{world6}},
\code{\link{world7}}
}
\concept{Core}
\keyword{variable}
