% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_496.R
\name{wornojob}
\alias{wornojob}
\title{Do you worry about not finding job}
\source{
General Social Survey https://gss.norc.org
}
\description{
wornojob
}
\details{
Question To what extent, if at all, do you worry about the possibility of not finding a job?
Would you say...
}
\section{Values}{

\itemize{
\item \verb{[1]} i worry a great deal
\item \verb{[2]} i worry to some extent
\item \verb{[3]} i worry a little
\item \verb{[4]} i don't worry at all
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6342/vshow}{GSS Data Explorer page for \code{wornojob}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab i worry a great deal \tab i worry to some extent \tab i worry a little \tab i don't worry at all \tab don't know \tab no answer \tab Total \cr
   2016 \tab 2533 \tab 49 \tab 36 \tab 71 \tab 153 \tab 1 \tab 24 \tab 2867 \cr
   Total \tab 2533 \tab 49 \tab 36 \tab 71 \tab 153 \tab 1 \tab 24 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{exjobsat}},
\code{\link{extraern}},
\code{\link{extrawrk}},
\code{\link{jbendmo}},
\code{\link{jbendyr}},
\code{\link{lowpay}},
\code{\link{mortravl}},
\code{\link{newskill}},
\code{\link{tempwork}}
}
\concept{Work Orientation}
\keyword{variable}
