% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_505.R
\name{worryjob}
\alias{worryjob}
\title{Do you worry about losing job}
\source{
General Social Survey https://gss.norc.org
}
\description{
worryjob
}
\details{
Question To what extent, if at all, do you worry about the possibility of losing your job?
}
\section{Values}{

\itemize{
\item \code{1} i worry a great deal
\item \code{2} i worry to some extent
\item \code{3} i worry a little
\item \code{4} i don't worry at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1998 \tab A/B/C \tab full \cr
   2006 \tab D/-/- \tab partial \cr
   2016 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4592/vshow}{GSS Data Explorer page for \code{worryjob}}.
}

\seealso{
Other Work Orientation: 
\code{\link{evjob}},
\code{\link{getjob}},
\code{\link{leavejob}},
\code{\link{lookjob}},
\code{\link{priagncy}},
\code{\link{pubagncy}},
\code{\link{wantjob2}},
\code{\link{whyjbend}},
\code{\link{yrjobend}}
}
\concept{Work Orientation}
\keyword{variable}
