% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_505.R
\name{worryjob}
\alias{worryjob}
\title{Do you worry about losing job}
\source{
General Social Survey https://gss.norc.org
}
\description{
worryjob
}
\details{
Question To what extent, if at all, do you worry about the possibility of losing your job?
}
\section{Values}{

\itemize{
\item \code{1} i worry a great deal
\item \code{2} i worry to some extent
\item \code{3} i worry a little
\item \code{4} i don't worry at all
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4592/vshow}{GSS Data Explorer page for \code{worryjob}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab i worry a great deal \tab i worry to some extent \tab i worry a little \tab i don't worry at all \tab don't know \tab no answer \tab Total \cr
   1998 \tab 1967 \tab 30 \tab 105 \tab 232 \tab 464 \tab 1 \tab 33 \tab 2832 \cr
   2006 \tab 3492 \tab 48 \tab 141 \tab 264 \tab 563 \tab - \tab 2 \tab 4510 \cr
   2016 \tab 1927 \tab 53 \tab 98 \tab 256 \tab 525 \tab 3 \tab 5 \tab 2867 \cr
   Total \tab 7386 \tab 131 \tab 344 \tab 752 \tab 1552 \tab 4 \tab 40 \tab 10209 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{evjob}},
\code{\link{getjob}},
\code{\link{leavejob}},
\code{\link{lookjob}},
\code{\link{priagncy}},
\code{\link{pubagncy}},
\code{\link{wantjob2}},
\code{\link{whyjbend}},
\code{\link{yrjobend}}
}
\concept{Work Orientation}
\keyword{variable}
