% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_449.R
\name{wotrial}
\alias{wotrial}
\title{Should authorities have right to detain the suspected}
\source{
General Social Survey https://gss.norc.org
}
\description{
wotrial
}
\details{
Question Suppose the government suspected that a terrorist act was about to happen. Do you think the authorities should have the right to…
A. Detain people for as long as they want without putting them on trial?
}
\section{Values}{

\itemize{
\item \verb{[1]} definitely should have right
\item \verb{[2]} probably should have right
\item \verb{[3]} probably should not have right
\item \verb{[4]} definitely should not have right
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B \tab 2016 \cr
   D/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4114/vshow}{GSS Data Explorer page for \code{wotrial}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab definitely should have right \tab probably should have right \tab probably should not have right \tab definitely should not have right \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 2992 \tab 351 \tab 449 \tab 347 \tab 348 \tab 21 \tab 2 \tab 4510 \cr
   2016 \tab 1477 \tab 205 \tab 290 \tab 391 \tab 468 \tab 23 \tab 13 \tab 2867 \cr
   Total \tab 4469 \tab 556 \tab 739 \tab 738 \tab 816 \tab 44 \tab 15 \tab 7377 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{exploit}},
\code{\link{fewtrsty}},
\code{\link{givinffor}},
\code{\link{givinfusa}},
\code{\link{hlpinflu}},
\code{\link{polsfair}},
\code{\link{stoprndm}},
\code{\link{tapphone}},
\code{\link{youinflu}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
