% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_01.R
\name{wrkgovt}
\alias{wrkgovt}
\title{Govt or private employee}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkgovt
}
\details{
Question A. What kind of work do you (did you normally) do?  That is, what (is/was) your job called?
OCCUPATION:  (See REMARKS)
B. What (do/did) you actually do in that job?  Tell me, what (are/were) some of your main duties?
C. What kind of place (do/did) you work for?
INDUSTRY:  (See REMARKS)
D. What (do/did) they (make/do)?
E. (Are/Were) you self employed or (do/did) you work for someone else?
F. (Are/Were) you employed by the federal, state or local government or by a private employer (including non-profit organizations)?
G. About how much time (does/did) it usually take you to travel to work - about how many minutes?
}
\section{Values}{

\itemize{
\item \code{1} government
\item \code{2} private
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1985 \tab -/-/-/- \tab full \cr
   1986 \tab -/-/-/- \tab full \cr
   2000 \tab A/B/C/- \tab full \cr
   2002 \tab A/B/C/- \tab full \cr
   2004 \tab A/B/C/- \tab full \cr
   2006 \tab A/B/C/D \tab full \cr
   2008 \tab A/B/C/- \tab full \cr
   2010 \tab A/B/C/- \tab full \cr
   2012 \tab A/B/C/- \tab full \cr
   2014 \tab A/B/C/- \tab full \cr
   2016 \tab A/B/C/- \tab full \cr
   2018 \tab A/B/C/- \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/10/vshow}{GSS Data Explorer page for \code{wrkgovt}}.
}

\seealso{
Other Core: 
\code{\link{evwork}},
\code{\link{hrs1}},
\code{\link{hrs2}},
\code{\link{id}},
\code{\link{occ}},
\code{\link{prestige}},
\code{\link{wrkslf}},
\code{\link{wrkstat}},
\code{\link{year}}
}
\concept{Core}
\keyword{variable}
