% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_623.R
\name{wrkgovt1}
\alias{wrkgovt1}
\title{Govt employee}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkgovt1
}
\details{
Question Are/were you employed by the government? (Please consider federal, state, or local government.)
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7432/vshow}{GSS Data Explorer page for \code{wrkgovt1}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 79 \tab 869 \tab 3047 \tab 2 \tab 5 \tab 30 \tab 4032 \cr
   2022 \tab 113 \tab 658 \tab 2746 \tab 16 \tab 4 \tab 7 \tab 3544 \cr
   Total \tab 192 \tab 1527 \tab 5793 \tab 18 \tab 9 \tab 37 \tab 7576 \cr
}
}

\keyword{variable}
