% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_623.R
\name{wrkgovt2}
\alias{wrkgovt2}
\title{Private employee}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkgovt2
}
\details{
Question Are/were you employed by a private employer (including non-profit organizations)?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7433/vshow}{GSS Data Explorer page for \code{wrkgovt2}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2021 \tab 79 \tab 2322 \tab 1597 \tab 2 \tab 5 \tab 27 \tab 4032 \cr
   2022 \tab 113 \tab 2157 \tab 1221 \tab 35 \tab 4 \tab 14 \tab 3544 \cr
   Total \tab 192 \tab 4479 \tab 2818 \tab 37 \tab 9 \tab 41 \tab 7576 \cr
}
}

\keyword{variable}
