% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_478.R
\name{wrkgrown}
\alias{wrkgrown}
\title{Shld woman work after kids leave home?}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkgrown
}
\details{
Question Do you think that women should work outside the home full-time, part-time or not at all under these circumstances:
D. After the children leave home.
}
\section{Values}{

\itemize{
\item \code{1} work full-time
\item \code{2} work part-time
\item \code{3} stay home
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1988 \tab A/B/C \tab full \cr
   1994 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4356/vshow}{GSS Data Explorer page for \code{wrkgrown}}.
}

\seealso{
Other ISSP Family & Gender Roles: 
\code{\link{daycare1}},
\code{\link{daycare2}},
\code{\link{daycare3}},
\code{\link{daycare4}},
\code{\link{daycare5}},
\code{\link{ilikejob}},
\code{\link{wrkbaby}},
\code{\link{wrknokid}},
\code{\link{wrksch}}
}
\concept{ISSP Family & Gender Roles}
\keyword{variable}
