% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_284.R
\name{wrkhro}
\alias{wrkhro}
\title{Hours per week using other computer make money}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkhro
}
\details{
Question For about how many minutes or hours a week do you use (this other computer/these other computers) for each of the following:
C. Paid work or other income-producing activities?
2. Hours per week?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2590/vshow}{GSS Data Explorer page for \code{wrkhro}}.
}

\seealso{
Other Information Society: 
\code{\link{emget}},
\code{\link{emgetp}},
\code{\link{emhro}},
\code{\link{emmino}},
\code{\link{emsent}},
\code{\link{emsentp}},
\code{\link{wrkmino}},
\code{\link{wwwhro}},
\code{\link{wwwmino}}

Other Computers And Internet: 
\code{\link{emget}},
\code{\link{emgetp}},
\code{\link{emhro}},
\code{\link{emmino}},
\code{\link{emsent}},
\code{\link{emsentp}},
\code{\link{wrkmino}},
\code{\link{wwwhro}},
\code{\link{wwwmino}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
