% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_492.R
\name{wrkimp}
\alias{wrkimp}
\title{Work is persons most impt. activity}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkimp
}
\details{
Question Thinking of work in general, please circle one number for each statement below to show how much you agree or disagree with each.
C. Work is a person's most important activity
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} neither
\item \code{4} disagree
\item \code{5} strongly disagree
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4494/vshow}{GSS Data Explorer page for \code{wrkimp}}.

Counts by year:\tabular{lllllllllll}{
   year \tab not available in this year \tab strongly agree \tab agree \tab neither \tab disagree \tab strongly disagree \tab can't choose \tab no answer \tab iap \tab Total \cr
   1989 \tab - \tab 180 \tab 344 \tab 299 \tab 458 \tab 137 \tab 14 \tab 105 \tab - \tab 1537 \cr
   1998 \tab - \tab 93 \tab 263 \tab 225 \tab 428 \tab 176 \tab 15 \tab 187 \tab 1445 \tab 2832 \cr
   Total \tab 0 \tab 273 \tab 607 \tab 524 \tab 886 \tab 313 \tab 29 \tab 292 \tab 1445 \tab 4369 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{ageemp}},
\code{\link{dowell}},
\code{\link{expernc}},
\code{\link{hwduties}},
\code{\link{paysame}},
\code{\link{sexemp}},
\code{\link{wrkearn}},
\code{\link{wrkenjoy}},
\code{\link{yrsfirm}}
}
\concept{Work Orientation}
\keyword{variable}
