% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_494.R
\name{wrkindp}
\alias{wrkindp}
\title{Importance of independent work in a job}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkindp
}
\details{
Question On the following list there are various aspects of jobs. Please circle one number to show how important you personally consider it is in a job:
F. A job that allows someone to work independently?
}
\section{Values}{

\itemize{
\item \verb{[1]} very important
\item \verb{[2]} important
\item \verb{[3]} neither important nor unimportant
\item \verb{[4]} not important
\item \verb{[5]} not important at all
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4512/vshow}{GSS Data Explorer page for \code{wrkindp}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab very important \tab important \tab neither important nor unimportant \tab not important \tab not important at all \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 84 \tab 380 \tab 660 \tab 279 \tab 81 \tab 10 \tab 19 \tab 24 \tab 1537 \cr
   1998 \tab 1609 \tab 335 \tab 586 \tab 198 \tab 35 \tab 6 \tab 23 \tab 40 \tab 2832 \cr
   2006 \tab 2992 \tab 544 \tab 655 \tab 228 \tab 73 \tab 11 \tab 5 \tab 2 \tab 4510 \cr
   2016 \tab 1390 \tab 423 \tab 665 \tab 319 \tab 57 \tab 4 \tab 4 \tab 5 \tab 2867 \cr
   Total \tab 6075 \tab 1682 \tab 2566 \tab 1024 \tab 246 \tab 31 \tab 51 \tab 71 \tab 11746 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{discwk5}},
\code{\link{famorjob}},
\code{\link{flexhrs}},
\code{\link{flexhrs1}},
\code{\link{hlpoths}},
\code{\link{hlpsoc}},
\code{\link{joborfam}},
\code{\link{whydisc5}},
\code{\link{wkpersnl}}
}
\concept{Work Orientation}
\keyword{variable}
