% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_652.R
\name{wrkmeangfl}
\alias{wrkmeangfl}
\title{I find my work meaningful}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkmeangfl
}
\details{
Question (Please tell us whether you strongly agree, agree, disagree, or strongly disagree with each of these statements.) I find my work to be very meaningful.
}
\section{Values}{

\itemize{
\item \code{1} strongly agree
\item \code{2} agree
\item \code{3} disagree
\item \code{4} strongly disagree
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7855/vshow}{GSS Data Explorer page for \code{wrkmeangfl}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab strongly agree \tab agree \tab disagree \tab strongly disagree \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 1548 \tab 680 \tab 1034 \tab 186 \tab 53 \tab 12 \tab 11 \tab 20 \tab 3544 \cr
   Total \tab 1548 \tab 680 \tab 1034 \tab 186 \tab 53 \tab 12 \tab 11 \tab 20 \tab 3544 \cr
}
}

\keyword{variable}
