% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_143.R
\name{wrkmuch}
\alias{wrkmuch}
\title{How much you compare to people do same job}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkmuch
}
\details{
Question Compared to other people who do the same or similar kind of work that you do, how much work would you say you do?  Would you say that you do much more, somewhat more, about the same, somewhat less or much less?
}
\section{Values}{

\itemize{
\item \code{1} much more
\item \code{2} somewhat more
\item \code{3} about same
\item \code{4} somewhat less
\item \code{5} much less
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1339/vshow}{GSS Data Explorer page for \code{wrkmuch}}.
}

\seealso{
Other Work Organizations: 
\code{\link{chngeorg}},
\code{\link{helporg}},
\code{\link{imatter}},
\code{\link{noticed}},
\code{\link{notloyal}},
\code{\link{proudorg}},
\code{\link{samevals}},
\code{\link{stayorg1}},
\code{\link{stayorg2}}
}
\concept{Work Organizations}
\keyword{variable}
