% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_495.R
\name{wrkshift}
\alias{wrkshift}
\title{What is your working schedule}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkshift
}
\details{
Question (HAND CARD D10)
Which of the following statements best describes your usual working schedule in your main job?
}
\section{Values}{

\itemize{
\item \code{1} i have a regular schedule or shift (daytime, evening, or night)
\item \code{2} i have a schedule or shift which regularly changes (for example, from days to evenings or to nights)
\item \code{3} i have a schedule where daily working times are decided at short notice by my employer
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2016 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/6333/vshow}{GSS Data Explorer page for \code{wrkshift}}.
}

\seealso{
Other Work Orientation: 
\code{\link{chngewrk}},
\code{\link{decidwrk}},
\code{\link{harass5}},
\code{\link{nounemp5}},
\code{\link{nounemp6}},
\code{\link{rcontact}},
\code{\link{unionsbd}},
\code{\link{wkathome}},
\code{\link{wkwkends}}
}
\concept{Work Orientation}
\keyword{variable}
