% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_495.R
\name{wrkshift}
\alias{wrkshift}
\title{What is your working schedule}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkshift
}
\details{
Question (HAND CARD D10)
Which of the following statements best describes your usual working schedule in your main job?
}
\section{Values}{

\itemize{
\item \verb{[1]} i have a regular schedule or shift (daytime, evening, or night)
\item \verb{[2]} i have a schedule or shift which regularly changes (for example, from days to evenings or to nights)
\item \verb{[3]} i have a schedule where daily working times are decided at short notice by my employer
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2016 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/6333/vshow}{GSS Data Explorer page for \code{wrkshift}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab i have a regular schedule or shift (daytime, evening, or night) \tab i have a schedule or shift which regularly changes (for example, from days to evenings or to nights) \tab i have a schedule where daily working times are decided at short notice by my employer \tab don't know \tab no answer \tab Total \cr
   2016 \tab 1927 \tab 700 \tab 166 \tab 55 \tab 11 \tab 8 \tab 2867 \cr
   Total \tab 1927 \tab 700 \tab 166 \tab 55 \tab 11 \tab 8 \tab 2867 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{chngewrk}},
\code{\link{decidwrk}},
\code{\link{harass5}},
\code{\link{nounemp5}},
\code{\link{nounemp6}},
\code{\link{rcontact}},
\code{\link{unionsbd}},
\code{\link{wkathome}},
\code{\link{wkwkends}}
}
\concept{Work Orientation}
\keyword{variable}
