% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_337.R
\name{wrkslffam}
\alias{wrkslffam}
\title{Does r work in family business or farm}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkslffam
}
\details{
Question Do/did) you work in your own family business or farm?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2018, 2021, 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7224/vshow}{GSS Data Explorer page for \code{wrkslffam}}.

Counts by year:\tabular{llllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab skipped on web \tab Total \cr
   2018 \tab 2109 \tab 130 \tab 102 \tab 7 \tab - \tab - \tab 2348 \cr
   2021 \tab 3574 \tab 193 \tab 236 \tab 5 \tab 1 \tab 23 \tab 4032 \cr
   2022 \tab 3129 \tab 177 \tab 212 \tab 7 \tab 3 \tab 16 \tab 3544 \cr
   Total \tab 8812 \tab 500 \tab 550 \tab 19 \tab 4 \tab 39 \tab 9924 \cr
}
}

\keyword{variable}
