% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_304.R
\name{wrktype}
\alias{wrktype}
\title{Work arrangement at main job}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrktype
}
\details{
Question How would you describe your work arrangement in your main job?
}
\section{Values}{

\itemize{
\item \code{1} independent contractor/consultant/freelance worker
\item \code{2} on-call, work only when called to work
\item \code{3} paid by a temporary agency
\item \code{4} work for contractor who provides workers/services
\item \code{5} regular, permanent employee
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2010 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2760/vshow}{GSS Data Explorer page for \code{wrktype}}.
}

\seealso{
Other Quality of Working Life: 
\code{\link{chngtme}},
\code{\link{moredays}},
\code{\link{mustwork}},
\code{\link{waypaid}},
\code{\link{wrksched}},
\code{\link{yearsjob}}
}
\concept{Quality of Working Life}
\keyword{variable}
