% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_136.R
\name{wrkyears}
\alias{wrkyears}
\title{#Years you have worked for pay since were 16}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrkyears
}
\details{
Question Since you were 16 years old, about how many years have you worked for pay?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   1991 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/1271/vshow}{GSS Data Explorer page for \code{wrkyears}}.
}

\seealso{
Other Work Organizations: 
\code{\link{empyears}},
\code{\link{evunemp}},
\code{\link{fndjob1}},
\code{\link{fndjob2}},
\code{\link{fndjob3}},
\code{\link{fndjob4}},
\code{\link{numunemp}},
\code{\link{orgfin}},
\code{\link{totunemp}}
}
\concept{Work Organizations}
\keyword{variable}
