% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_532.R
\name{wrldgovt}
\alias{wrldgovt}
\title{International bodies should enforce environment}
\source{
General Social Survey https://gss.norc.org
}
\description{
wrldgovt
}
\details{
Question How much do you agree or disagree with the following statements?
B. For certain problems, like environmental pollution, international bodies should have the right to enforce solutions
}
\section{Values}{

\itemize{
\item \code{1} agree strongly
\item \code{2} agree
\item \code{3} neither agree nor disagree
\item \code{4} disagree
\item \code{5} disagree strongly
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1996, 2004, 2014 \cr
   B/C/- \tab 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4851/vshow}{GSS Data Explorer page for \code{wrldgovt}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab agree strongly \tab agree \tab neither agree nor disagree \tab disagree \tab disagree strongly \tab can't choose \tab no answer \tab Total \cr
   1996 \tab 1537 \tab 262 \tab 650 \tab 199 \tab 116 \tab 28 \tab 85 \tab 27 \tab 2904 \cr
   2004 \tab 1596 \tab 225 \tab 501 \tab 245 \tab 141 \tab 56 \tab 47 \tab 1 \tab 2812 \cr
   2014 \tab 1264 \tab 164 \tab 564 \tab 232 \tab 150 \tab 65 \tab 95 \tab 4 \tab 2538 \cr
   Total \tab 4397 \tab 651 \tab 1715 \tab 676 \tab 407 \tab 149 \tab 227 \tab 32 \tab 8254 \cr
}
}

\seealso{
Other ISSP National Identity: 
\code{\link{amcult}},
\code{\link{amownway}},
\code{\link{amtv}},
\code{\link{forland}},
\code{\link{forlang}},
\code{\link{imports}},
\code{\link{mincult}},
\code{\link{proudgrp}},
\code{\link{proudhis}}
}
\concept{ISSP National Identity}
\keyword{variable}
