% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_670.R
\name{wtssnrps_nea}
\alias{wtssnrps_nea}
\title{Poststratification and nonresponse weight for analyzing nea follow-on}
\source{
General Social Survey https://gss.norc.org
}
\description{
wtssnrps_nea
}
\details{
Question POSTSTRATIFICATION AND NONRESPONSE WEIGHT FOR ANALYZING NEA FOLLOW-ON
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\keyword{variable}
