% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_271.R
\name{wwwgthic}
\alias{wwwgthic}
\title{In last month listen gothic music on web}
\source{
General Social Survey https://gss.norc.org
}
\description{
wwwgthic
}
\details{
Question In the last month, how many times have you used the Web to learn about or listen to the following kinds of music?
C. Gothic or Industrial?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} 1-2 times
\item \code{3} more than 2 times
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2463/vshow}{GSS Data Explorer page for \code{wwwgthic}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab never \tab 1-2 times \tab more than 2 times \tab no answer \tab Total \cr
   2000 \tab 2498 \tab 185 \tab 15 \tab 6 \tab 113 \tab 2817 \cr
   Total \tab 2498 \tab 185 \tab 15 \tab 6 \tab 113 \tab 2817 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{musicbuy}},
\code{\link{musiclst}},
\code{\link{wwwclass}},
\code{\link{wwwcntry}},
\code{\link{wwwjazz}},
\code{\link{wwwoldie}},
\code{\link{wwwrap}},
\code{\link{wwwrelig}},
\code{\link{wwwrock}}
}
\concept{Information Society}
\keyword{variable}
