% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_263.R
\name{wwwhr}
\alias{wwwhr}
\title{Www hours per week}
\source{
General Social Survey https://gss.norc.org
}
\description{
wwwhr
}
\details{
Question Not counting e-mail, about how many minutes or hours per week Do you use the Web? (Include time you spend visiting regular web sites and time spent using interactive Internet services like chat rooms, Usenet groups, discussion forums, bulletin boards, and the like.)
A. Hours per week?
}
\section{Values}{

\itemize{
\item \code{0} 0 hours
\item \code{168} 168 hours
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
   2004 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2010 \tab A/B/- \tab partial \cr
   2012 \tab A/B/- \tab partial \cr
   2014 \tab A/B/- \tab partial \cr
   2016 \tab A/B/- \tab partial \cr
   2018 \tab A/B/- \tab partial \cr
   2021 \tab A/B/- \tab partial \cr
   2022 \tab A/B/- \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2384/vshow}{GSS Data Explorer page for \code{wwwhr}}.
}

\seealso{
Other Information Society: 
\code{\link{chathr}},
\code{\link{chatmin}},
\code{\link{drctlink}},
\code{\link{emailhr}},
\code{\link{emailmin}},
\code{\link{srcheng}},
\code{\link{typeurl}},
\code{\link{usewww}},
\code{\link{wwwmin}}
}
\concept{Information Society}
\keyword{variable}
