% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_280.R
\name{wwwline2}
\alias{wwwline2}
\title{Means to www}
\source{
General Social Survey https://gss.norc.org
}
\description{
wwwline2
}
\details{
Question D. What is this other means by which you are connected to the World Wide Web?
}
\section{Values}{

\itemize{
\item \code{1} special high-speed tele line (isdn) cncted to cmptr
\item \code{2} cable service line cnctd to computer
\item \code{3} web tv line, cnctd to tv set
\item \code{4} more than one type of cnction
\item \code{5} other, not named above
\item \code{6} wireless connection (e.g. from a satellite)
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/- \tab partial \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2555/vshow}{GSS Data Explorer page for \code{wwwline2}}.
}

\seealso{
Other Information Society: 
\code{\link{comphome}},
\code{\link{ecomhrh}},
\code{\link{ecomminh}},
\code{\link{numhome}},
\code{\link{paywww}},
\code{\link{perminh}},
\code{\link{usehome}},
\code{\link{wwwhome}},
\code{\link{wwwline1}}

Other Computers And Internet: 
\code{\link{comphome}},
\code{\link{ecomhrh}},
\code{\link{ecomminh}},
\code{\link{numhome}},
\code{\link{paywww}},
\code{\link{perminh}},
\code{\link{usehome}},
\code{\link{wwwhome}},
\code{\link{wwwline1}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
