% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_273.R
\name{wwwlit}
\alias{wwwlit}
\title{Use web to learn about literature}
\source{
General Social Survey https://gss.norc.org
}
\description{
wwwlit
}
\details{
Question A. Do you ever use the Web to learn about, read, or discuss topics related to literature–novels, poetry, or plays?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2479/vshow}{GSS Data Explorer page for \code{wwwlit}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2326 \tab 148 \tab 228 \tab 115 \tab 2817 \cr
   2002 \tab 2284 \tab 164 \tab 271 \tab 46 \tab 2765 \cr
   Total \tab 4610 \tab 312 \tab 499 \tab 161 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{litauth}},
\code{\link{litget}},
\code{\link{litread}},
\code{\link{litsite}},
\code{\link{wwwbiz}},
\code{\link{wwwcowrk}},
\code{\link{wwwgfbf}},
\code{\link{wwwpersn}},
\code{\link{wwwsp}}
}
\concept{Information Society}
\keyword{variable}
