% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_274.R
\name{wwwnei}
\alias{wwwnei}
\title{Has anyone met on wen become neighbor}
\source{
General Social Survey https://gss.norc.org
}
\description{
wwwnei
}
\details{
Question F. Neighbor?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2489/vshow}{GSS Data Explorer page for \code{wwwnei}}.

Counts by year:\tabular{llllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab Total \cr
   2000 \tab 2657 \tab 4 \tab 65 \tab 91 \tab 2817 \cr
   2002 \tab 2647 \tab 4 \tab 67 \tab 47 \tab 2765 \cr
   Total \tab 5304 \tab 8 \tab 132 \tab 138 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{askother}},
\code{\link{askwork}},
\code{\link{callcomp}},
\code{\link{doonown}},
\code{\link{manual}},
\code{\link{online}},
\code{\link{payother}},
\code{\link{wwwfrnd}},
\code{\link{wwwvol}}
}
\concept{Information Society}
\keyword{variable}
