% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_273.R
\name{wwwpersn}
\alias{wwwpersn}
\title{Has r met person on web}
\source{
General Social Survey https://gss.norc.org
}
\description{
wwwpersn
}
\details{
Question A. Have you ever first come into contact with someone on the Internet–through a chat room, bulletin board, news group, discussion forum, or other interactive site, or through their Web page—with whom you later established a relationship outside the Internet?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2000, 2002 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/2484/vshow}{GSS Data Explorer page for \code{wwwpersn}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab no answer \tab don't know \tab Total \cr
   2000 \tab 2075 \tab 70 \tab 582 \tab 90 \tab - \tab 2817 \cr
   2002 \tab 2043 \tab 71 \tab 603 \tab 47 \tab 1 \tab 2765 \cr
   Total \tab 4118 \tab 141 \tab 1185 \tab 137 \tab 1 \tab 5582 \cr
}
}

\seealso{
Other Information Society: 
\code{\link{litauth}},
\code{\link{litget}},
\code{\link{litread}},
\code{\link{litsite}},
\code{\link{wwwbiz}},
\code{\link{wwwcowrk}},
\code{\link{wwwgfbf}},
\code{\link{wwwlit}},
\code{\link{wwwsp}}
}
\concept{Information Society}
\keyword{variable}
