% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_274.R
\name{wwwvol}
\alias{wwwvol}
\title{Has anyone met on web become volunteer with r}
\source{
General Social Survey https://gss.norc.org
}
\description{
wwwvol
}
\details{
Question H. Someone you worked with on a civic, community or political project?
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2000 \tab A/B/C \tab full \cr
   2002 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/2491/vshow}{GSS Data Explorer page for \code{wwwvol}}.
}

\seealso{
Other Information Society: 
\code{\link{askother}},
\code{\link{askwork}},
\code{\link{callcomp}},
\code{\link{doonown}},
\code{\link{manual}},
\code{\link{online}},
\code{\link{payother}},
\code{\link{wwwfrnd}},
\code{\link{wwwnei}}

Other Computers And Internet: 
\code{\link{askother}},
\code{\link{askwork}},
\code{\link{callcomp}},
\code{\link{doonown}},
\code{\link{manual}},
\code{\link{online}},
\code{\link{payother}},
\code{\link{wwwfrnd}},
\code{\link{wwwnei}}
}
\concept{Computers And Internet}
\concept{Information Society}
\keyword{variable}
