% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_500.R
\name{xhaustn}
\alias{xhaustn}
\title{How often does work exhaust r}
\source{
General Social Survey https://gss.norc.org
}
\description{
xhaustn
}
\details{
Question Now some more questions about your working conditions. Please circle one code for each item below to show how often it applies to your work. How often...
A. Do you come home from work exhausted?
}
\section{Values}{

\itemize{
\item \code{1} always
\item \code{2} often
\item \code{3} sometimes
\item \code{4} hardly ever
\item \code{5} never
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1989, 1998 \cr
   B/C/- \tab 2016 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4550/vshow}{GSS Data Explorer page for \code{xhaustn}}.

Counts by year:\tabular{llllllllll}{
   year \tab iap \tab always \tab often \tab sometimes \tab hardly ever \tab never \tab can't choose \tab no answer \tab Total \cr
   1989 \tab 639 \tab 55 \tab 259 \tab 441 \tab 96 \tab 20 \tab 3 \tab 24 \tab 1537 \cr
   1998 \tab 1967 \tab 68 \tab 274 \tab 398 \tab 71 \tab 17 \tab 3 \tab 34 \tab 2832 \cr
   2006 \tab 3492 \tab 151 \tab 251 \tab 465 \tab 116 \tab 33 \tab - \tab 2 \tab 4510 \cr
   2016 \tab 1927 \tab 150 \tab 294 \tab 367 \tab 101 \tab 21 \tab 1 \tab 6 \tab 2867 \cr
   Total \tab 8025 \tab 424 \tab 1078 \tab 1671 \tab 384 \tab 91 \tab 7 \tab 66 \tab 11746 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{physwrk}},
\code{\link{rflexhrs}},
\code{\link{rhiinc}},
\code{\link{rhlpoths}},
\code{\link{rhlpsoc}},
\code{\link{rintjob}},
\code{\link{rleisure}},
\code{\link{rpromotn}},
\code{\link{rwrkindp}}
}
\concept{Work Orientation}
\keyword{variable}
