% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_71.R
\name{xmovie1}
\alias{xmovie1}
\title{Seen porn film (theater or vcr) last yr}
\source{
General Social Survey https://gss.norc.org
}
\description{
xmovie1
}
\details{
Question B. Have you seen an X-rated movie either in a movie theater or on a VCR (Video Cassette Recorder) in the last year?
}
\section{Values}{

\itemize{
\item \verb{[1]} yes
\item \verb{[2]} no
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 1994 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/642/vshow}{GSS Data Explorer page for \code{xmovie1}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab yes \tab no \tab don't know \tab no answer \tab Total \cr
   1994 \tab 2486 \tab 113 \tab 390 \tab 1 \tab 2 \tab 2992 \cr
   Total \tab 2486 \tab 113 \tab 390 \tab 1 \tab 2 \tab 2992 \cr
}
}

\seealso{
Other Core: 
\code{\link{pornfirm}},
\code{\link{pornimp}},
\code{\link{porninfo}},
\code{\link{pornlaw}},
\code{\link{pornmorl}},
\code{\link{pornout}},
\code{\link{pornrape}},
\code{\link{spanking}},
\code{\link{xmovie}}
}
\concept{Core}
\keyword{variable}
