% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_643.R
\name{xmoviey}
\alias{xmoviey}
\title{Seen pornography in past year}
\source{
General Social Survey https://gss.norc.org
}
\description{
xmoviey
}
\details{
Question The next questions are about pornography- Internet materials or websites, books, movies, magazines, or photographs that explicitly show sex activities. About how often did you look at pornography in the last year?
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} less than once a year
\item \verb{[3]} about once or twice a year
\item \verb{[4]} several times a year
\item \verb{[5]} about once a month
\item \verb{[6]} 2-3 times a month
\item \verb{[7]} nearly every week
\item \verb{[8]} every week
\item \verb{[9]} once a day
\item \verb{[10]} several times a day
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   B/C \tab 2022, 2024 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7757/vshow}{GSS Data Explorer page for \code{xmoviey}}.

Counts by year:\tabular{llllllllllllllll}{
   year \tab iap \tab never \tab less than once a year \tab about once or twice a year \tab several times a year \tab about once a month \tab 2-3 times a month \tab nearly every week \tab every week \tab once a day \tab several times a day \tab don't know \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2364 \tab 627 \tab 79 \tab 84 \tab 97 \tab 47 \tab 55 \tab 40 \tab 47 \tab 24 \tab 12 \tab 39 \tab 25 \tab 4 \tab 3544 \cr
   Total \tab 2364 \tab 627 \tab 79 \tab 84 \tab 97 \tab 47 \tab 55 \tab 40 \tab 47 \tab 24 \tab 12 \tab 39 \tab 25 \tab 4 \tab 3544 \cr
}
}

\keyword{variable}
