% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_643.R
\name{xmoviey}
\alias{xmoviey}
\title{Seen pornography in past year}
\source{
General Social Survey https://gss.norc.org
}
\description{
xmoviey
}
\details{
Question The next questions are about pornography- Internet materials or websites, books, movies, magazines, or photographs that explicitly show sex activities. About how often did you look at pornography in the last year?
}
\section{Values}{

\itemize{
\item \code{1} never
\item \code{2} less than once a year
\item \code{3} about once or twice a year
\item \code{4} several times a year
\item \code{5} about once a month
\item \code{6} 2-3 times a month
\item \code{7} nearly every week
\item \code{8} every week
\item \code{9} once a day
\item \code{10} several times a day
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab B/C \tab partial \cr
   2024 \tab B/C \tab partial \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7757/vshow}{GSS Data Explorer page for \code{xmoviey}}.
}

\keyword{variable}
