% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_652.R
\name{yearval1}
\alias{yearval1}
\title{Estimate of dollar value of payments in that year}
\source{
General Social Survey https://gss.norc.org
}
\description{
yearval1
}
\details{
Question NA
}
\section{Values}{

\itemize{
\item \code{1} less than $100
\item \code{2} $100 to $299
\item \code{3} $300 to $499
\item \code{4} $500 to $999
\item \code{5} $1,000 to $1,499
\item \code{6} $1,500 to $2,499
\item \code{7} $2,500 to $3,999
\item \code{8} $4,000 to $6,999
\item \code{9} $7,000 to $14,999
\item \code{10} $15,000 to $49,999
\item \code{11} $50,000 to $99,999
\item \code{12} over $100,000
\item \code{NA(d)} don't know
\item \code{NA(i)} iap
\item \code{NA(j)} I don't have a job
\item \code{NA(m)} dk, na, iap
\item \code{NA(n)} no answer
\item \code{NA(p)} not imputable
\item \code{NA(r)} refused
\item \code{NA(s)} skipped on web
\item \code{NA(u)} uncodeable
\item \code{NA(x)} not available in this release
\item \code{NA(y)} not available in this year
\item \code{NA(z)} see codebook
}
}

\keyword{variable}
