% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_652.R
\name{yearval1}
\alias{yearval1}
\title{Estimate of dollar value of payments in that year}
\source{
General Social Survey https://gss.norc.org
}
\description{
yearval1
}
\details{
Question What was the approximate total dollar value of these various payments in that year? NOTE: All responses who give valid answers to EXTRAYR are asked YEARVAL.
}
\section{Values}{

\itemize{
\item \code{1} less than $100
\item \code{2} $100 to $299
\item \code{3} $300 to $499
\item \code{4} $500 to $999
\item \code{5} $1,000 to $1,499
\item \code{6} $1,500 to $2,499
\item \code{7} $2,500 to $3,999
\item \code{8} $4,000 to $6,999
\item \code{9} $7,000 to $14,999
\item \code{10} $15,000 to $49,999
\item \code{11} $50,000 to $99,999
\item \code{12} over $100,000
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2002 \tab A/B/C \tab full \cr
   2006 \tab A/B/C \tab full \cr
   2014 \tab A/B/C \tab full \cr
   2018 \tab A/B/C \tab full \cr
   2022 \tab B/C/- \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/8039/vshow}{GSS Data Explorer page for \code{yearval1}}.
}

\seealso{
Other Employee Compensation: 
\code{\link{compperf}},
\code{\link{cowrkhrd}},
\code{\link{deptperf}},
\code{\link{donothng}},
\code{\link{extr2001}},
\code{\link{extrapay}},
\code{\link{extraval}},
\code{\link{extrayr}},
\code{\link{indperf}},
\code{\link{lastwkmo}},
\code{\link{lastwkyr}},
\code{\link{ownstock}},
\code{\link{seecowrk}},
\code{\link{stockops}},
\code{\link{stockval}},
\code{\link{talkemp}},
\code{\link{talksup}},
\code{\link{workfor}},
\code{\link{workfor1}},
\code{\link{yearval}}
}
\concept{Employee Compensation}
\keyword{variable}
