% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_449.R
\name{youinflu}
\alias{youinflu}
\title{How often r is asked to help influence important decision}
\source{
General Social Survey https://gss.norc.org
}
\description{
youinflu
}
\details{
Question Some people because of their job, position in the community or contacts, are asked by others to  help influence important decisions in their favor.
A. What about you? How often are you asked to help influence important decisions in other people's favor?
}
\section{Values}{

\itemize{
\item \verb{[1]} never
\item \verb{[2]} seldom
\item \verb{[3]} occasionally
\item \verb{[4]} often
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   D \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4119/vshow}{GSS Data Explorer page for \code{youinflu}}.

Counts by year:\tabular{lllllllll}{
   year \tab iap \tab never \tab seldom \tab occasionally \tab often \tab can't choose \tab no answer \tab Total \cr
   2006 \tab 2992 \tab 544 \tab 445 \tab 396 \tab 127 \tab 5 \tab 1 \tab 4510 \cr
   Total \tab 2992 \tab 544 \tab 445 \tab 396 \tab 127 \tab 5 \tab 1 \tab 4510 \cr
}
}

\seealso{
Other ISSP Role of Gov't: 
\code{\link{exploit}},
\code{\link{fewtrsty}},
\code{\link{givinffor}},
\code{\link{givinfusa}},
\code{\link{hlpinflu}},
\code{\link{polsfair}},
\code{\link{stoprndm}},
\code{\link{tapphone}},
\code{\link{wotrial}}
}
\concept{ISSP Role of Gov't}
\keyword{variable}
