% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_58.R
\name{youngen}
\alias{youngen}
\title{What should younger generation be taught}
\source{
General Social Survey https://gss.norc.org
}
\description{
youngen
}
\details{
Question Here are two statements. Will you tell me which one you agree with more?
A. First, the younger generation should be taught by their elders to do what is right.
B. Second, the younger generation should be taught to think for themselves even though they may do something their elders disapprove of.
}
\section{Values}{

\itemize{
\item \verb{[1]} what elders think
\item \verb{[2]} think for selves
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   - \tab 1980 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/520/vshow}{GSS Data Explorer page for \code{youngen}}.

Counts by year:\tabular{lllllll}{
   year \tab iap \tab what elders think \tab think for selves \tab don't know \tab no answer \tab Total \cr
   1980 \tab - \tab 658 \tab 797 \tab 6 \tab 7 \tab 1468 \cr
   Total \tab 0 \tab 658 \tab 797 \tab 6 \tab 7 \tab 1468 \cr
}
}

\seealso{
Other Core: 
\code{\link{chldsex}},
\code{\link{chldsex1}},
\code{\link{helpoth}},
\code{\link{socbar}},
\code{\link{socfrend}},
\code{\link{socommun}},
\code{\link{socrel}},
\code{\link{thnkself}},
\code{\link{workhard}}
}
\concept{Core}
\keyword{variable}
