% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_645.R
\name{yrfnce}
\alias{yrfnce}
\title{Financial situation compared to last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrfnce
}
\details{
Question Compared to 12 months ago, how do you think the financial situation of your household has changed? Would you say the current situation is...
}
\section{Values}{

\itemize{
\item \verb{[1]} much better
\item \verb{[2]} somewhat better
\item \verb{[3]} about the same
\item \verb{[4]} somewhat worse
\item \verb{[5]} much worse
\item \verb{[NA(d)]} can't choose \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A \tab 2022 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7787/vshow}{GSS Data Explorer page for \code{yrfnce}}.

Counts by year:\tabular{lllllllllll}{
   year \tab iap \tab much better \tab somewhat better \tab about the same \tab somewhat worse \tab much worse \tab can't choose \tab no answer \tab skipped on web \tab Total \cr
   2022 \tab 2386 \tab 73 \tab 211 \tab 456 \tab 280 \tab 91 \tab 34 \tab 8 \tab 5 \tab 3544 \cr
   Total \tab 2386 \tab 73 \tab 211 \tab 456 \tab 280 \tab 91 \tab 34 \tab 8 \tab 5 \tab 3544 \cr
}
}

\keyword{variable}
