% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_646.R
\name{yrfnce}
\alias{yrfnce}
\title{Financial situation compared to last year}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrfnce
}
\details{
Question Compared to 12 months ago, how do you think the financial situation of your household has changed? Would you say the current situation is...
}
\section{Values}{

\itemize{
\item \code{1} much better
\item \code{2} somewhat better
\item \code{3} about the same
\item \code{4} somewhat worse
\item \code{5} much worse
\item \code{NA(d)} can't choose / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/7787/vshow}{GSS Data Explorer page for \code{yrfnce}}.
}

\keyword{variable}
