% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_357.R
\name{yrinfrm3}
\alias{yrinfrm3}
\title{How many years r born a member of informal group3}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrinfrm3
}
\details{
Question Q. How many years have you been a member of…?
3. Third informal group
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3218/vshow}{GSS Data Explorer page for \code{yrinfrm3}}.

Counts by year:\tabular{llllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 13 \tab 15 \tab 2 \tab 20 \tab 26 \tab 3 \tab 4 \tab 40 \tab 44 \tab 5 \tab 6 \tab 7 \tab 8 \tab no answer \tab Total \cr
   2004 \tab 2767 \tab 2 \tab 6 \tab 2 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 4 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 13 \tab 2812 \cr
   Total \tab 2767 \tab 2 \tab 6 \tab 2 \tab 1 \tab 1 \tab 2 \tab 3 \tab 1 \tab 4 \tab 3 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 13 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtfrat1}},
\code{\link{mtfrat2}},
\code{\link{mtfrat3}},
\code{\link{yrinfrm1}},
\code{\link{yrinfrm2}},
\code{\link{yrinfrm4}},
\code{\link{yrinfrm5}},
\code{\link{yrother4}},
\code{\link{yrother5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
