% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_357.R
\name{yrinfrm3}
\alias{yrinfrm3}
\title{How many years r born a member of informal group3}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrinfrm3
}
\details{
Question Q. How many years have you been a member of…?
3. Third informal group
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3218/vshow}{GSS Data Explorer page for \code{yrinfrm3}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{mtfrat1}},
\code{\link{mtfrat2}},
\code{\link{mtfrat3}},
\code{\link{yrinfrm1}},
\code{\link{yrinfrm2}},
\code{\link{yrinfrm4}},
\code{\link{yrinfrm5}},
\code{\link{yrother4}},
\code{\link{yrother5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
