% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_505.R
\name{yrjobend}
\alias{yrjobend}
\title{When did your last paid job end}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrjobend
}
\details{
Question When did your last paid job end?
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 1998 \cr
   D/-/- \tab 2006 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/4594/vshow}{GSS Data Explorer page for \code{yrjobend}}.

Counts by year:\tabular{llllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 1941 \tab 1944 \tab 1945 \tab 1946 \tab 1950 \tab 1951 \tab 1952 \tab 1955 \tab 1956 \tab 1959 \tab 1962 \tab 1964 \tab 1966 \tab 1967 \tab 1969 \tab 1970 \tab 1971 \tab 1972 \tab 1974 \tab 1975 \tab 1976 \tab 1978 \tab 1979 \tab 1980 \tab 1981 \tab 1982 \tab 1983 \tab 1984 \tab 1985 \tab 1986 \tab 1987 \tab 1988 \tab 1989 \tab 1990 \tab 1991 \tab 1992 \tab 1993 \tab 1994 \tab 1995 \tab 1996 \tab 1997 \tab 1998 \tab no answer \tab 1926 \tab 1942 \tab 1954 \tab 1960 \tab 1961 \tab 1965 \tab 1973 \tab 1977 \tab 1999 \tab 2000 \tab 2001 \tab 2002 \tab 2003 \tab 2004 \tab 2005 \tab 2006 \tab Total \cr
   1998 \tab 2523 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 1 \tab 3 \tab 1 \tab 2 \tab 1 \tab 2 \tab 4 \tab 1 \tab 2 \tab 2 \tab 2 \tab 4 \tab 5 \tab 5 \tab 6 \tab 3 \tab 5 \tab 4 \tab 8 \tab 12 \tab 5 \tab 4 \tab 2 \tab 9 \tab 5 \tab 15 \tab 11 \tab 7 \tab 13 \tab 15 \tab 26 \tab 38 \tab 13 \tab 62 \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab - \tab 2832 \cr
   2006 \tab 4066 \tab - \tab - \tab - \tab - \tab - \tab 1 \tab - \tab 2 \tab 1 \tab - \tab 1 \tab - \tab 2 \tab - \tab 2 \tab 2 \tab - \tab 2 \tab 1 \tab 1 \tab 4 \tab 3 \tab 1 \tab 6 \tab 2 \tab 6 \tab 7 \tab 5 \tab 7 \tab 11 \tab 3 \tab 13 \tab 12 \tab 11 \tab 4 \tab 10 \tab 12 \tab 9 \tab 6 \tab 15 \tab 8 \tab 10 \tab 10 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 14 \tab 18 \tab 24 \tab 13 \tab 26 \tab 47 \tab 65 \tab 37 \tab 4510 \cr
   Total \tab 6589 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 3 \tab 2 \tab 1 \tab 4 \tab 1 \tab 4 \tab 1 \tab 4 \tab 6 \tab 1 \tab 4 \tab 3 \tab 3 \tab 8 \tab 8 \tab 6 \tab 12 \tab 5 \tab 11 \tab 11 \tab 13 \tab 19 \tab 16 \tab 7 \tab 15 \tab 21 \tab 16 \tab 19 \tab 21 \tab 19 \tab 22 \tab 21 \tab 41 \tab 46 \tab 23 \tab 72 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 2 \tab 1 \tab 1 \tab 14 \tab 18 \tab 24 \tab 13 \tab 26 \tab 47 \tab 65 \tab 37 \tab 7342 \cr
}
}

\seealso{
Other Work Orientation: 
\code{\link{evjob}},
\code{\link{getjob}},
\code{\link{leavejob}},
\code{\link{lookjob}},
\code{\link{priagncy}},
\code{\link{pubagncy}},
\code{\link{wantjob2}},
\code{\link{whyjbend}},
\code{\link{worryjob}}
}
\concept{Work Orientation}
\keyword{variable}
