% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_654.R
\name{yrlvmus}
\alias{yrlvmus}
\title{Attend live music or theater in past 12 months}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrlvmus
}
\details{
Question First, we'd like to ask about some leisure or recreational activities that people do during their free time. As we list each activity, please indicate if it is something you have done in the past 12 months. During the last 12 months, did you go in person to a live music, theater, or dance performance?

NOTE: This variable originates on the NEA follow-on, and should not be analyzed without the NEA-specific weights WTSSPS_NEA or WTSSNRPS_NEA
}
\section{Values}{

\itemize{
\item \code{1} yes
\item \code{2} no
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2022 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/7866/vshow}{GSS Data Explorer page for \code{yrlvmus}}.
}

\seealso{
Other NEA: 
\code{\link{artsout}},
\code{\link{yrartxbt}},
\code{\link{yrcreat}},
\code{\link{yrmovie}},
\code{\link{yrrdg}},
\code{\link{yrstmus}},
\code{\link{yrtour}}
}
\concept{NEA}
\keyword{variable}
