% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_351.R
\name{yrpolit2}
\alias{yrpolit2}
\title{How many years r been a member of political club2}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrpolit2
}
\details{
Question D. How many years have you been a member of…?
2. Second political club
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B \tab partial \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3162/vshow}{GSS Data Explorer page for \code{yrpolit2}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrpolit1}},
\code{\link{yrpolit3}},
\code{\link{yrserv7}},
\code{\link{yrserv8}},
\code{\link{yrvet1}},
\code{\link{yrvet2}},
\code{\link{yrvet3}},
\code{\link{yrvet4}},
\code{\link{yrvet5}}

Other Voluntary Associations: 
\code{\link{yrpolit1}},
\code{\link{yrpolit3}},
\code{\link{yrserv7}},
\code{\link{yrserv8}},
\code{\link{yrvet1}},
\code{\link{yrvet2}},
\code{\link{yrvet3}},
\code{\link{yrvet4}},
\code{\link{yrvet5}}
}
\concept{Voluntary Assocations/Social Networks}
\concept{Voluntary Associations}
\keyword{variable}
