% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_353.R
\name{yrschl1}
\alias{yrschl1}
\title{How many years r born a member of school service group1}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrschl1
}
\details{
Question H. How many years have you been a member of
\enumerate{
\item First school service group
}
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{lll}{
   year \tab ballots \tab availability \cr
   2004 \tab A/B/C \tab full \cr
}
}

\section{Link at the GSS}{

For details see the \href{https://gssdataexplorer.norc.org/variables/3179/vshow}{GSS Data Explorer page for \code{yrschl1}}.
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrhobby1}},
\code{\link{yrschl2}},
\code{\link{yrschl3}},
\code{\link{yrschl4}},
\code{\link{yryouth1}},
\code{\link{yryouth2}},
\code{\link{yryouth3}},
\code{\link{yryouth4}},
\code{\link{yryouth5}}

Other Voluntary Associations: 
\code{\link{yrhobby1}},
\code{\link{yrschl2}},
\code{\link{yrschl3}},
\code{\link{yrschl4}},
\code{\link{yryouth1}},
\code{\link{yryouth2}},
\code{\link{yryouth3}},
\code{\link{yryouth4}},
\code{\link{yryouth5}}
}
\concept{Voluntary Assocations/Social Networks}
\concept{Voluntary Associations}
\keyword{variable}
