% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_353.R
\name{yrschl2}
\alias{yrschl2}
\title{How many years r born a member of school service group2}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrschl2
}
\details{
Question H. How many years have you been a member of
2. Second school service group
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3180/vshow}{GSS Data Explorer page for \code{yrschl2}}.

Counts by year:\tabular{lllllllllllllll}{
   year \tab iap \tab 1 \tab 10 \tab 12 \tab 2 \tab 23 \tab 3 \tab 4 \tab 5 \tab 6 \tab 8 \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2785 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 9 \tab 2812 \cr
   Total \tab 2785 \tab 4 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 1 \tab 3 \tab 1 \tab 1 \tab 1 \tab 9 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrhobby1}},
\code{\link{yrschl1}},
\code{\link{yrschl3}},
\code{\link{yrschl4}},
\code{\link{yryouth1}},
\code{\link{yryouth2}},
\code{\link{yryouth3}},
\code{\link{yryouth4}},
\code{\link{yryouth5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
