% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_366.R
\name{yrskwn2}
\alias{yrskwn2}
\title{Years r has known the named2}
\source{
General Social Survey https://gss.norc.org
}
\description{
yrskwn2
}
\details{
Question How long have you known (NAME)?
B. How many years have you known the second name?
}
\section{Values}{

\itemize{
\item 
\item \code{NA(d)} don't know / \code{NA(i)} iap / \code{NA(j)} I don't have a job / \code{NA(m)} dk, na, iap / \code{NA(n)} no answer / \code{NA(p)} not imputable / \code{NA(r)} refused / \code{NA(s)} skipped on web / \code{NA(u)} uncodeable / \code{NA(x)} not available in this release / \code{NA(y)} not available in this year / \code{NA(z)} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3309/vshow}{GSS Data Explorer page for \code{yrskwn2}}.

Counts by year:\tabular{lllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 13 \tab 14 \tab 15 \tab 16 \tab 17 \tab 18 \tab 19 \tab 2 \tab 20 \tab 21 \tab 22 \tab 23 \tab 24 \tab 25 \tab 26 \tab 27 \tab 28 \tab 29 \tab 3 \tab 30 \tab 31 \tab 32 \tab 33 \tab 34 \tab 35 \tab 36 \tab 37 \tab 38 \tab 39 \tab 4 \tab 40 \tab 41 \tab 42 \tab 43 \tab 44 \tab 45 \tab 46 \tab 47 \tab 48 \tab 49 \tab 5 \tab 50 \tab 51 \tab 52 \tab 53 \tab 54 \tab 55 \tab 56 \tab 57 \tab 58 \tab 59 \tab 6 \tab 60 \tab 61 \tab 62 \tab 63 \tab 64 \tab 65 \tab 66 \tab 67 \tab 7 \tab 70 \tab 71 \tab 73 \tab 78 \tab 79 \tab 8 \tab 80 \tab 86 \tab 9 \tab no answer \tab Total \cr
   2004 \tab 1977 \tab 7 \tab 35 \tab 26 \tab 6 \tab 18 \tab 8 \tab 12 \tab 22 \tab 5 \tab 5 \tab 8 \tab 8 \tab 28 \tab 25 \tab 8 \tab 5 \tab 13 \tab 8 \tab 37 \tab 12 \tab 13 \tab 11 \tab 7 \tab 33 \tab 30 \tab 9 \tab 19 \tab 13 \tab 16 \tab 12 \tab 7 \tab 7 \tab 7 \tab 9 \tab 34 \tab 19 \tab 9 \tab 11 \tab 7 \tab 7 \tab 18 \tab 3 \tab 7 \tab 7 \tab 8 \tab 29 \tab 13 \tab 6 \tab 7 \tab 3 \tab 2 \tab 4 \tab 2 \tab 2 \tab 3 \tab 2 \tab 26 \tab 4 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 16 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 25 \tab 1 \tab 1 \tab 5 \tab 48 \tab 2812 \cr
   Total \tab 1977 \tab 7 \tab 35 \tab 26 \tab 6 \tab 18 \tab 8 \tab 12 \tab 22 \tab 5 \tab 5 \tab 8 \tab 8 \tab 28 \tab 25 \tab 8 \tab 5 \tab 13 \tab 8 \tab 37 \tab 12 \tab 13 \tab 11 \tab 7 \tab 33 \tab 30 \tab 9 \tab 19 \tab 13 \tab 16 \tab 12 \tab 7 \tab 7 \tab 7 \tab 9 \tab 34 \tab 19 \tab 9 \tab 11 \tab 7 \tab 7 \tab 18 \tab 3 \tab 7 \tab 7 \tab 8 \tab 29 \tab 13 \tab 6 \tab 7 \tab 3 \tab 2 \tab 4 \tab 2 \tab 2 \tab 3 \tab 2 \tab 26 \tab 4 \tab 2 \tab 2 \tab 1 \tab 1 \tab 1 \tab 2 \tab 2 \tab 16 \tab 2 \tab 1 \tab 1 \tab 1 \tab 1 \tab 25 \tab 1 \tab 1 \tab 5 \tab 48 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{frstmet2}},
\code{\link{frstmet3}},
\code{\link{frstmet4}},
\code{\link{frstmet5}},
\code{\link{relneg}},
\code{\link{yrskwn1}},
\code{\link{yrskwn3}},
\code{\link{yrskwn4}},
\code{\link{yrskwn5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
