% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gss_vars_353.R
\name{yryouth1}
\alias{yryouth1}
\title{How many years r born a member of youth groups1}
\source{
General Social Survey https://gss.norc.org
}
\description{
yryouth1
}
\details{
Question G. How many years have you been a member of
\enumerate{
\item First youth group
}
}
\section{Values}{

\itemize{
\item \verb{[NA(d)]} don't know \verb{[NA(i)]} iap \verb{[NA(j)]} I don't have a job \verb{[NA(m)]} dk, na, iap \verb{[NA(n)]} no answer \verb{[NA(p)]} not imputable \verb{[NA(r)]} refused \verb{[NA(s)]} skipped on web \verb{[NA(u)]} uncodeable \verb{[NA(x)]} not available in this release \verb{[NA(y)]} not available in this year \verb{[NA(z)]} see codebook
}
}

\section{Question Years and Ballots}{
\tabular{ll}{
   ballots \tab years \cr
   A/B/C \tab 2004 \cr
}
}

\section{Overview}{

For further details see the \href{https://gssdataexplorer.norc.org/variables/3174/vshow}{GSS Data Explorer page for \code{yryouth1}}.

Counts by year:\tabular{llllllllllllllllllllllllll}{
   year \tab iap \tab 0 \tab 1 \tab 10 \tab 11 \tab 12 \tab 14 \tab 15 \tab 17 \tab 19 \tab 2 \tab 20 \tab 3 \tab 30 \tab 33 \tab 4 \tab 5 \tab 50 \tab 6 \tab 7 \tab 8 \tab 9 \tab don't know \tab no answer \tab Total \cr
   2004 \tab 2656 \tab 6 \tab 29 \tab 9 \tab 4 \tab 2 \tab 2 \tab 4 \tab 1 \tab 1 \tab 21 \tab 8 \tab 14 \tab 2 \tab 1 \tab 9 \tab 9 \tab 1 \tab 10 \tab 6 \tab 4 \tab 1 \tab 3 \tab 9 \tab 2812 \cr
   Total \tab 2656 \tab 6 \tab 29 \tab 9 \tab 4 \tab 2 \tab 2 \tab 4 \tab 1 \tab 1 \tab 21 \tab 8 \tab 14 \tab 2 \tab 1 \tab 9 \tab 9 \tab 1 \tab 10 \tab 6 \tab 4 \tab 1 \tab 3 \tab 9 \tab 2812 \cr
}
}

\seealso{
Other Voluntary Assocations/Social Networks: 
\code{\link{yrhobby1}},
\code{\link{yrschl1}},
\code{\link{yrschl2}},
\code{\link{yrschl3}},
\code{\link{yrschl4}},
\code{\link{yryouth2}},
\code{\link{yryouth3}},
\code{\link{yryouth4}},
\code{\link{yryouth5}}
}
\concept{Voluntary Assocations/Social Networks}
\keyword{variable}
